Virtual Class BernsteinFunction for Bernstein Functions

A virtual superclass for all implementations of the various classes of Bernstein functions which have a representaion $$ \psi(x) = a + b x + \int_0^\infty (1 - e^{-ux}) \nu(du) , x > 0 $$ for non-negative constants \(a, b \geq 0\), called killing rate and drift, and a Lévy measure \(\nu\) on \((0, \infty)\).

valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

uexIntensities(object, d, ...)

intensities(object, d, ...)

exIntensities(object, d, ...)

exQMatrix(object, d, ...)

# S4 method for BernsteinFunction
uexIntensities(object, d, ...)

# S4 method for BernsteinFunction
exIntensities(object, d, ...)

# S4 method for BernsteinFunction
intensities(object, d, ...)

# S4 method for BernsteinFunction
exQMatrix(object, d, ...)

Arguments

object

An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

x

Non-negativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

The order of the alternating iterated forward differences taken on the Bernstein function (\(j\) in the representation).

n, k

Non-negative numbers for the binomial factors.

cscale

Positive number, the composit scaling factor.

...

Pass-through parameter

d

Dimension

Details

Bernstein functions have a 1-1 relationship to Lévy subordinators s.t. for every Bernstein function \(\psi\), there exists a Lévy subordinator \(\Lambda\) with $$ E[e^{-x \Lambda_t}] = e^{-t \psi(x)} , t, x > 0 . $$

\(a\) Describes the (exponential) killing rate, which determines the rate of an exponentially distributed event that sends the subordinator to infinity.

\(b\) Describes the linear drift of the Lévy subordinator.

\(\nu\) determines the arrival-rate and shock-size distribution for the jumps of the Lévy subordinator.

For a theoretic treatment of Bernstein function, we refer to (Schilling et al. 2012) .

The method valueOf is implemented different, depending on the specific Bernstein function to get the best possible accuracy:

For difference_order == 0L, the values are calculated with closed form formulas if possible.

For linear Bernstein functions and constant Bernstein functions the iterated differences are calculated automatically.

For scaled Bernstein functions and sums of Bernstein functions the result is calculted recursively.

For all other cases, the values are calculated using either the Lévy-Khintchine or the Stieltjes representation. Which one is chosen, depends on the case and can be manually overwritten with method = "levy" or method = "stieltjes".

Functions

valueOf: Calculate the values for a compound scaled Bernstein function and its higher-order, alternating iterated forward differences, possibly scaled by a binomial $$ {(-1)}^{j-1} \Delta^j \psi(c x), x > 0. $$

uexIntensities: Calculates (unscaled) exchangeable shock arrival intensities, see rmo() and rexmo().

intensities: Calculates intensities parameter for rmo().

exIntensities: Calculates (scaled) exchangeable shock-size arrival intensities for rexmo().

exQMatrix: Calculates the infinitesimal generator matrix of the corresponding Markovian default-counting process used internally by rexmo().

References

Schilling RL, Song R, Vondracek Z (2012). Bernstein functions, 2 edition. De Gruyter. doi: 10.1515/9783110269338 .

See also

Examples

exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.6970130 0.2615107 0.5933218

exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.6970130 0.2615107 0.5933218

intensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.23233766 0.23233766 0.08717025 0.23233766 0.08717025 0.08717025 0.59332184

exQMatrix(AlphaStableBernsteinFunction(4e-1), 3L)
#>           [,1]      [,2]       [,3]      [,4]
#> [1,] -1.551846  0.697013  0.2615107 0.5933218
#> [2,]  0.000000 -1.319508  0.6390158 0.6804921
#> [3,]  0.000000  0.000000 -1.0000000 1.0000000
#> [4,]  0.000000  0.000000  0.0000000 0.0000000