Virtual Class CompleteBernsteinFunction for Complete Bernstein Functions

A virtual superclass for all Bernstein functions which can representated by a Stieltjes density (no drift or killing rate). That means that there exists a Stieltjes measure \(\sigma\) such that $$ \psi(x) = \int_0^\infty \frac{x}{x + u} \sigma(du) , x > 0 . $$

stieltjesDensity(object)

# S4 method for CompleteBernsteinFunction
valueOf(
  object,
  x,
  difference_order,
  cscale = 1,
  n = 1L,
  k = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

Arguments

object
An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

x
Non-negativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order
The order of the alternating iterated forward differences taken on the Bernstein function (\(j\) in the representation).

cscale
Positive number, the composit scaling factor.

n
Non-negative numbers for the binomial factors.

k
Non-negative numbers for the binomial factors.

...
Further parameter (passed to stats::integrate())

method
Method to calculate the result; use method = "levy" for using the Lévy representation and method = "stieltjes" for using the Stieltjes representation.

tolerance
(Relative) tolerance, passed down to stats::integrate()

Details

For continuous Stieltjes densities, the values of the Bernstein function are calculated with stats::integrate() by using the representation $$ \psi(x) = \int_{0}^{\infty} x \mathrm{Beta}(1, x + u) \sigma(du), \quad x > 0 , $$ and the values of the iterated differences are calculated by using the representation $$ (-1)^{j-1} \Delta^{j} \psi(x) = \int_{0}^{\infty} u \mathrm{Beta}(j+1, x + u) \sigma(du) , \quad x > 0 . $$

For discrete Lévy densities \(\sigma(du) = \sum_{i} y_i \delta_{u_i}(du)\), the values of the Bernstein function are calculated by using the representation $$ \psi(x) = \sum_{i} x \mathrm{Beta}(1, x + u_i) y_i, \quad x > 0 , $$ and the values of the iterated differences are calculated by using the representation $$ (-1)^{j-1} \Delta^{j} \psi(x) = \sum_{i} u_i \mathrm{Beta}(j+1, x + u_i) y_i , \quad x > 0 . $$

Functions

stieltjesDensity: returns the Stieltjes density with lower, upper, and type attributes if continuous and returns a data.frame with named columns x (atoms) and y (weights) as well as a type attribute if discrete. The type attribute is either "continuous" or "discrete".

valueOf,CompleteBernsteinFunction-method: Calculates the iterated differences of the Bernstein function, see valueOf()

See also