Class for composite scaled Bernstein functions

Bernstein functions are stable under composition, i.e. if \(\psi\) is a Bernstein function and c > 0, then $$ x \mapsto \psi(c x) $$ is also a Bernstein function.

Usage

# S4 method for CompositeScaledBernsteinFunction
initialize(.Object, cscale, original)

# S4 method for CompositeScaledBernsteinFunction
show(object)

# S4 method for CompositeScaledBernsteinFunction
valueOf(object, x, difference_order = 0L, n = 1, k = 0, cscale = 1, ...)

Arguments

.Object

An object: see the “Initialize Methods” section.

cscale

Positive number.

original

Derives from BernsteinFunction.

object

An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

x

a nonnegativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

a nonnegative integer with the order of the alternating iterated forward differences taken on the Bernstein function.

n, k

nonnegative numbers for the binomial factor.

...

arguments to specify properties of the new object, to be passed to initialize().

Functions

initialize(CompositeScaledBernsteinFunction): Constructor

show(CompositeScaledBernsteinFunction): Display the object.

valueOf(CompositeScaledBernsteinFunction): Calculates the iterated differences of the Bernstein function, see valueOf()

Slots

cscale

The scale of the inner linear Bernstein function of the composition.

original

The original Bernstein function.

See also

Examples

CompositeScaledBernsteinFunction()
#> An object of class "CompositeScaledBernsteinFunction"
#> 	 (invalid or not initialized)
cscale <- 0.5
bf_original <- AlphaStableBernsteinFunction()
CompositeScaledBernsteinFunction(cscale = cscale, original = bf_original)
#> An object of class "CompositeScaledBernsteinFunction"
#> - cscale: 0.5
#> - original:
#> 	An object of class "AlphaStableBernsteinFunction"
#> 		 (invalid or not initialized)