Class for the Gamma Bernstein function

The Gamma Bernstein function, is the Bernstein function of a subordinator with a (scaled) Gamma distribution. The representation is for \(a > 0\) $$ \psi(x) = \log(1 + \frac{x}{a}), x > 0. $$

# S4 method for GammaBernsteinFunction
initialize(.Object, a = 1)

# S4 method for GammaBernsteinFunction
levyDensity(object)

# S4 method for GammaBernsteinFunction
stieltjesDensity(object)

Arguments

.Object
An object: see the “Initialize Methods” section.

a
Positive number.

object
An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

Details

For this Bernstein function, the higher-order alternating iterated forward differences are known in closed form but cannot be evaluated numerically without the danger of loss of significance. But we can use numerical integration (here: stats::integrate()) to approximate it with the following representation: $$ {(-1)}^{k-1} \Delta^{k} \psi(x) = \int_{0}^{\infty} e^{-ux} {(1 - e^{-u})}^{k} \frac{e^{-au}}{u} du, x>0, k>0. $$

This Bernstein function is no. 26 in the list of complete Bernstein functions in Chp. 16 of (Schilling et al. 2012) .

Functions

initialize,GammaBernsteinFunction-method: Constructor

levyDensity,GammaBernsteinFunction-method: see LevyBernsteinFunction

stieltjesDensity,GammaBernsteinFunction-method: see CompleteBernsteinFunction

Slots

a

Scale parameter for the Lévy measure.

Lévy density

$$ \nu(du) = \frac{\operatorname{e}^{-a u}}{u}, \quad u > 0 . $$

Stieltjes density

$$ \sigma(du) = 1 / u du, u > a . $$

References

Schilling RL, Song R, Vondracek Z (2012). Bernstein functions, 2 edition. De Gruyter. doi: 10.1515/9783110269338 , https://doi.org/10.1515/9783110269338.

See also

Examples

GammaBernsteinFunction()

#> An object of class "GammaBernsteinFunction"
#> Slot "a":
#> [1] 1
#> 

GammaBernsteinFunction(a = 2)

#> An object of class "GammaBernsteinFunction"
#> Slot "a":
#> [1] 2
#>