Class for sums of two Bernstein functions

Bernstein functions are stable under addition, i.e. if \(\psi_1\) and \(\psi_2\) are two Bernstein functions, then $$ x \mapsto \psi_1(x) + \psi_2(x) , x>0, $$ is also a Bernstein function.

Usage

# S4 method for SumOfBernsteinFunctions
initialize(.Object, first, second)

# S4 method for SumOfBernsteinFunctions
show(object)

# S4 method for SumOfBernsteinFunctions
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

Arguments

.Object

An object: see the “Initialize Methods” section.

first

Derives from BernsteinFunction.

second

Derives from BernsteinFunction.

object

An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

x

a non-negativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

a non-negative integer with the order of the alternating iterated forward differences taken on the Bernstein function.

n, k

non-negative numbers for the binomial factor.

cscale

a positive numeric scalar with the composite scaling factor.

...

arguments to specify properties of the new object, to be passed to initialize().

Functions

initialize(SumOfBernsteinFunctions): Constructor

show(SumOfBernsteinFunctions): Display the object.

valueOf(SumOfBernsteinFunctions): Calculates the iterated differences of the Bernstein function, see valueOf()

Slots

first

The first summand (derived from BernsteinFunction).

second

The second summand (derived from BernsteinFunction).

See also

Examples

SumOfBernsteinFunctions()
#> An object of class "SumOfBernsteinFunctions"
#> 	 (invalid or not initialized)
first_bf <- LinearBernsteinFunction(scale = 0.2)
second_bf <- AlphaStableBernsteinFunction(alpha = 0.5)
SumOfBernsteinFunctions(first = first_bf, second = second_bf)
#> An object of class "SumOfBernsteinFunctions"
#> - first:
#> 	An object of class "LinearBernsteinFunction"
#> 	- scale: 0.2
#> - second:
#> 	An object of class "AlphaStableBernsteinFunction"
#> 	- alpha: 0.5