For S4 methods that require a documentation entry but only clutter the index.

For S4 methods that require a documentation entry but only clutter the index.

Usage

# S4 method for BernsteinFunction
uexIntensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
exIntensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
intensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
exQMatrix(object, d, cscale = 1, ...)

# S4 method for LevyBernsteinFunction
defaultMethod(object)

# S4 method for LevyBernsteinFunction
valueOf(
  object,
  x,
  difference_order,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...,
  method = c("default", "levy"),
  tolerance = .Machine$double.eps^0.5
)

# S4 method for LevyBernsteinFunction
valueOf0(object, x, cscale = 1, method = defaultMethod(object), ...)

# S4 method for CompleteBernsteinFunction
valueOf(
  object,
  x,
  difference_order,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

# S4 method for AlphaStableBernsteinFunction
initialize(.Object, alpha)

# S4 method for AlphaStableBernsteinFunction
show(object)

# S4 method for AlphaStableBernsteinFunction
levyDensity(object)

# S4 method for AlphaStableBernsteinFunction
stieltjesDensity(object)

# S4 method for AlphaStableBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for CompositeScaledBernsteinFunction
initialize(.Object, cscale, original)

# S4 method for CompositeScaledBernsteinFunction
show(object)

# S4 method for CompositeScaledBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for CompositeScaledBernsteinFunction
valueOf(object, x, difference_order = 0L, n = 1, k = 0, cscale = 1, ...)

# S4 method for ConstantBernsteinFunction
initialize(.Object, constant)

# S4 method for ConstantBernsteinFunction
show(object)

# S4 method for ConstantBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for ConstantBernsteinFunction
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

# S4 method for ConvexCombinationOfBernsteinFunctions
initialize(.Object, coefficients, points)

# S4 method for ConvexCombinationOfBernsteinFunctions
show(object)

# S4 method for ConvexCombinationOfBernsteinFunctions
valueOf0(object, x, cscale = 1, ...)

# S4 method for ConvexCombinationOfBernsteinFunctions
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

# S4 method for ExponentialBernsteinFunction
initialize(.Object, lambda)

# S4 method for ExponentialBernsteinFunction
show(object)

# S4 method for ExponentialBernsteinFunction
levyDensity(object)

# S4 method for ExponentialBernsteinFunction
stieltjesDensity(object)

# S4 method for ExponentialBernsteinFunction
defaultMethod(object)

# S4 method for ExponentialBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for GammaBernsteinFunction
initialize(.Object, a)

# S4 method for GammaBernsteinFunction
show(object)

# S4 method for GammaBernsteinFunction
levyDensity(object)

# S4 method for GammaBernsteinFunction
stieltjesDensity(object)

# S4 method for GammaBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for InverseGaussianBernsteinFunction
initialize(.Object, eta)

# S4 method for InverseGaussianBernsteinFunction
show(object)

# S4 method for InverseGaussianBernsteinFunction
levyDensity(object)

# S4 method for InverseGaussianBernsteinFunction
stieltjesDensity(object)

# S4 method for InverseGaussianBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for LinearBernsteinFunction
initialize(.Object, scale)

# S4 method for LinearBernsteinFunction
show(object)

# S4 method for LinearBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for LinearBernsteinFunction
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

# S4 method for ParetoBernsteinFunction
initialize(.Object, alpha, x0)

# S4 method for ParetoBernsteinFunction
show(object)

# S4 method for ParetoBernsteinFunction
levyDensity(object)

# S4 method for ParetoBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for PoissonBernsteinFunction
initialize(.Object, eta)

# S4 method for PoissonBernsteinFunction
show(object)

# S4 method for PoissonBernsteinFunction
levyDensity(object)

# S4 method for PoissonBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for ScaledBernsteinFunction
initialize(.Object, scale, original)

# S4 method for ScaledBernsteinFunction
show(object)

# S4 method for ScaledBernsteinFunction
valueOf0(object, x, cscale = 1, ...)

# S4 method for ScaledBernsteinFunction
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

# S4 method for SumOfBernsteinFunctions
initialize(.Object, first, second)

# S4 method for SumOfBernsteinFunctions
show(object)

# S4 method for SumOfBernsteinFunctions
valueOf0(object, x, cscale = 1, ...)

# S4 method for SumOfBernsteinFunctions
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

Arguments

object

An object deriving from the class BernsteinFunction.

d

A positive integer, larger than two, for the dimension.

cscale

Positive number.

...

pass-through parameter.

x

A nonnegative numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

A nonnegative integer with the order of the alternating iterated forward differences taken on the Bernstein function.

n, k

Nonnegative numbers for the binomial factor.

method

Method to calculate the result; use method = "levy" for using the Lévy representation and method = "stieltjes" for using the Stieltjes representation.

tolerance

(Relative) tolerance, passed down to stats::integrate().

.Object

An object: see the “Initialize Methods” section.

alpha

Positive number between zero and one (bounds excl.).

original

Derives from BernsteinFunction.

constant

Nonnegative number.

coefficients

Derives from ConvexCombinationOfBernsteinFunctions.

points

Derives from ConvexCombinationOfBernsteinFunctions.

lambda

Positive number.

a

Positive number.

eta

Positive number.

scale

Positive number.

x0

Positive number.

first

Derives from BernsteinFunction.

second

Derives from BernsteinFunction.