Sample from the armageddon ESM distribution

Draws n independent samples from a d variate armageddon ESM distribution with parameters alpha and beta.

rarmextmo_esm(n, d, alpha, beta)

Arguments

n
Number of samples

d
Dimension

alpha
Shock intensity of individual shocks

beta
Shock intensity of global shock

Value

rarmextmo_esm implements an optimized version of the exogenous shock model algorithm for the armageddon shock family and returns an \(n \times d\) array matrix with rows corresponding to the independent samples of size \(d\).

Details

Parameterisation: The armageddon ESM distribution is a special case of the exchangeable Marshall-Olkin distribution where \(\lambda_i = 0 \forall 1 < i < d\). The individual shock rate is \(\alpha = \lambda_1\) and the global shock rate is \(\beta = \lambda_d\).

See also

Examples

rarmextmo_esm(10L, 2L, 0.5, 0.2)
#>            [,1]       [,2]
#>  [1,] 1.3373321 0.40304354
#>  [2,] 8.2198095 1.70443232
#>  [3,] 1.0900860 0.04746857
#>  [4,] 2.6049203 3.90330608
#>  [5,] 4.1960497 2.49590487
#>  [6,] 1.0046901 0.07131968
#>  [7,] 0.7542240 1.95823879
#>  [8,] 0.2500026 0.72065167
#>  [9,] 1.0786992 0.53952187
#> [10,] 3.5674030 0.12629416
rarmextmo_esm(10L, 2L, 0, 1)      ## comonotone
#>            [,1]      [,2]
#>  [1,] 0.8346728 0.8346728
#>  [2,] 2.1228993 2.1228993
#>  [3,] 1.1099744 1.1099744
#>  [4,] 0.6503988 0.6503988
#>  [5,] 0.7884202 0.7884202
#>  [6,] 0.1400899 0.1400899
#>  [7,] 1.0360235 1.0360235
#>  [8,] 0.1925256 0.1925256
#>  [9,] 1.9184386 1.9184386
#> [10,] 0.3760188 0.3760188
rarmextmo_esm(10L, 2L, 1, 0)      ## independence
#>            [,1]       [,2]
#>  [1,] 0.8075358 0.71259757
#>  [2,] 0.7070553 1.20103089
#>  [3,] 0.1198257 0.26672977
#>  [4,] 0.2012750 0.73863261
#>  [5,] 0.8533478 0.97090525
#>  [6,] 1.4269226 0.10330351
#>  [7,] 1.7654759 3.79108831
#>  [8,] 0.6281026 1.30879119
#>  [9,] 1.6762804 0.01530056
#> [10,] 0.3207186 0.02358263