Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5866770 3.5421530 2.6123422
#>  [2,] 0.1991350 0.1991350 0.1937796
#>  [3,] 0.2985208 2.6370964 0.2985208
#>  [4,] 2.4423895 1.4482304 3.0836667
#>  [5,] 3.6588958 0.2647230 2.8259865
#>  [6,] 1.7910702 1.0906001 2.5254785
#>  [7,] 1.6403552 0.9620905 2.1651466
#>  [8,] 0.8125660 0.7950287 0.7328271
#>  [9,] 0.7569618 0.7569618 0.6994321
#> [10,] 5.3503835 0.3861681 0.3861681
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3926775 0.1477956 2.4661466
#>  [2,] 0.4334356 0.9546733 0.3307002
#>  [3,] 1.6966330 0.5913069 0.3280064
#>  [4,] 0.7438120 0.1539990 1.6663373
#>  [5,] 1.3135295 1.8586109 0.1022216
#>  [6,] 0.6473380 0.4949691 1.3598130
#>  [7,] 0.3589648 1.3307064 1.4467683
#>  [8,] 1.0295085 0.6672209 0.4965972
#>  [9,] 0.8986268 1.1426658 0.4851671
#> [10,] 0.4998578 0.8054762 0.1325618
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1238803 1.1238803 1.1238803
#>  [2,] 0.2138858 0.2138858 0.2138858
#>  [3,] 2.5356087 2.5356087 2.5356087
#>  [4,] 0.1665868 0.1665868 0.1665868
#>  [5,] 0.6895311 0.6895311 0.6895311
#>  [6,] 1.2249038 1.2249038 1.2249038
#>  [7,] 0.3805103 0.3805103 0.3805103
#>  [8,] 0.5749208 0.5749208 0.5749208
#>  [9,] 0.4408363 0.4408363 0.4408363
#> [10,] 1.2342241 1.2342241 1.2342241

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.77785191 2.32467337 0.006592321
#>  [2,] 0.43047074 0.43047074 0.430470744
#>  [3,] 1.05795308 0.30161527 1.057953077
#>  [4,] 0.75376346 0.75376346 0.753763462
#>  [5,] 0.04374768 0.14104755 0.043747677
#>  [6,] 2.74501579 0.27437326 0.274373255
#>  [7,] 0.18371357 0.12085593 0.120855934
#>  [8,] 0.80746660 0.26261467 0.807466599
#>  [9,] 0.07727528 0.07727528 0.077275277
#> [10,] 0.01549079 0.01549079 0.015490791
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.61148710 2.44859950 1.1131558
#>  [2,] 0.09527778 0.09529575 2.9542451
#>  [3,] 1.67107349 0.88176594 0.2734746
#>  [4,] 0.34126130 2.68679194 0.8445852
#>  [5,] 0.52109750 1.09097615 1.3042119
#>  [6,] 1.08690957 1.35421607 1.1361304
#>  [7,] 2.60760332 1.62835641 0.6294458
#>  [8,] 0.03588075 1.09673687 0.3537615
#>  [9,] 1.49511570 0.02322249 0.9617709
#> [10,] 0.18665752 0.11757934 0.3824106
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31679312 0.31679312 0.31679312
#>  [2,] 0.09604214 0.09604214 0.09604214
#>  [3,] 0.03849543 0.03849543 0.03849543
#>  [4,] 1.81722541 1.81722541 1.81722541
#>  [5,] 0.23250193 0.23250193 0.23250193
#>  [6,] 0.31291861 0.31291861 0.31291861
#>  [7,] 0.14420566 0.14420566 0.14420566
#>  [8,] 0.73764691 0.73764691 0.73764691
#>  [9,] 0.48814810 0.48814810 0.48814810
#> [10,] 0.11941251 0.11941251 0.11941251

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9899095 0.8500655 2.37133461
#>  [2,] 0.6003003 0.1842604 3.66092153
#>  [3,] 0.7791530 0.5483632 0.77915300
#>  [4,] 0.4715820 0.5331666 0.53316656
#>  [5,] 0.4384418 0.4317149 0.98713660
#>  [6,] 1.5768003 3.5082902 1.13057714
#>  [7,] 0.4901014 0.4901014 0.02503938
#>  [8,] 1.9563448 1.9563448 1.95634478
#>  [9,] 0.2659641 0.1810043 0.18100430
#> [10,] 1.6673033 0.9380188 1.85450574
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2468525 1.34612747 1.7946388
#>  [2,] 0.9843540 0.72276694 2.8753270
#>  [3,] 1.7866460 0.56419240 1.6111130
#>  [4,] 1.9134513 3.06709415 0.1300537
#>  [5,] 0.3899166 2.80807831 0.7841806
#>  [6,] 1.1408408 0.05032661 1.2467521
#>  [7,] 0.1269064 1.79523587 0.1767900
#>  [8,] 1.2052007 2.80691782 2.2586852
#>  [9,] 0.9614660 1.12684806 0.5537312
#> [10,] 0.2315398 0.06963431 0.5405783
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29126600 0.29126600 0.29126600
#>  [2,] 0.47842411 0.47842411 0.47842411
#>  [3,] 1.65253649 1.65253649 1.65253649
#>  [4,] 0.03355208 0.03355208 0.03355208
#>  [5,] 0.51246624 0.51246624 0.51246624
#>  [6,] 0.72866994 0.72866994 0.72866994
#>  [7,] 2.92037748 2.92037748 2.92037748
#>  [8,] 0.25677227 0.25677227 0.25677227
#>  [9,] 1.22322961 1.22322961 1.22322961
#> [10,] 1.86697986 1.86697986 1.86697986

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.34632732 0.81755687 0.18730321
#>  [2,] 0.11962984 0.11962984 0.11962984
#>  [3,] 0.69929824 0.88703057 1.22173428
#>  [4,] 0.02667971 0.02667971 0.02667971
#>  [5,] 0.20237068 0.03072825 0.20237068
#>  [6,] 1.21643005 0.05581362 1.21643005
#>  [7,] 0.20473250 0.16749613 0.20473250
#>  [8,] 0.33628587 0.33628587 0.33628587
#>  [9,] 0.39720857 0.46705142 0.39720857
#> [10,] 0.41565450 1.65069435 0.56302147
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.11012426 3.70236741 1.6331450
#>  [2,] 1.91739372 0.04961996 2.8483605
#>  [3,] 0.09711874 1.24429332 1.2777882
#>  [4,] 0.12902760 1.22900561 1.9409555
#>  [5,] 1.18122204 1.76427477 0.4164729
#>  [6,] 0.88530604 0.33564284 1.6657320
#>  [7,] 0.35543415 0.62575603 1.3863339
#>  [8,] 1.03595216 1.61990849 1.4238077
#>  [9,] 0.54876367 0.01280151 0.5703022
#> [10,] 1.34140595 0.05488392 0.4321956
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.82684497 0.82684497 0.82684497
#>  [2,] 0.02117104 0.02117104 0.02117104
#>  [3,] 0.33479069 0.33479069 0.33479069
#>  [4,] 2.77263116 2.77263116 2.77263116
#>  [5,] 0.19426782 0.19426782 0.19426782
#>  [6,] 0.70278181 0.70278181 0.70278181
#>  [7,] 0.09776131 0.09776131 0.09776131
#>  [8,] 1.76804032 1.76804032 1.76804032
#>  [9,] 1.42891176 1.42891176 1.42891176
#> [10,] 1.35626011 1.35626011 1.35626011