Sample from an exchangeable Marshall-Olkin distribution

Draws n independent samples from a d variate exchangeable Marshall-Olkin distribution with (scaled) shock rates ex_intensities.

rexmo_mdcm(n, d, ex_intensities)

Arguments

n
Number of samples

d
dimension

ex_intensities
(Scaled) exchangeable Marshall-Olkin intensities

Value

rexmo_mdcm implements the Markovian model for the default counting process of the exchangeable subclass and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically disctributed samples of a \(d\) variate exchangeable Marshall-Olkin distribution with exchangeable parameters ex_intensities.

Details

Parameterisation: The exchangeable Marshall--Olkin distribution has the property that \(\lambda_I\) only depends on the cardinality of \(I\). The unscaled exchangeable shock intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\). The scaled exchangeable shock intensities are defined by \(\eta_i = \binom{d}{i} \lambda_i\).

The scaled exchangeable shock intensities:

The scaled exchangeable shock intensities* \(\eta_i\) must be stored in a vector of length \(d\).

The entry \(\eta_i\) is the intensity of a shock corresponding to a set with \(i\) elements multiplied by the number of shocks of cardinality \(i\), i.e. \(\binom{d}{i}\).

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 122 psqq.

See also

Examples

rexmo_mdcm(10, 2, c(2 * 0.4, 0.2))
#>            [,1]      [,2]
#>  [1,] 0.6843666 0.6711049
#>  [2,] 1.1290153 1.4105994
#>  [3,] 0.4209118 0.1337887
#>  [4,] 2.1935123 2.8316439
#>  [5,] 6.8842395 3.2003378
#>  [6,] 1.7667651 1.7667651
#>  [7,] 4.5151703 1.9538939
#>  [8,] 2.3520736 0.1367053
#>  [9,] 0.8659696 0.1627321
#> [10,] 0.4435195 2.7592044
rexmo_mdcm(10, 2, c(2, 0))      ## independence
#>             [,1]      [,2]
#>  [1,] 1.26595820 0.1678158
#>  [2,] 1.29523247 1.2215184
#>  [3,] 0.84763997 1.8282054
#>  [4,] 0.32801442 0.5937467
#>  [5,] 0.01112249 1.4248105
#>  [6,] 0.28112257 1.1572701
#>  [7,] 1.62138527 0.2591584
#>  [8,] 0.46997429 0.1559826
#>  [9,] 0.36668246 0.2151861
#> [10,] 0.09847499 0.1338374
rexmo_mdcm(10, 2, c(0, 1))      ## comonotone
#>             [,1]       [,2]
#>  [1,] 0.34689565 0.34689565
#>  [2,] 0.35595619 0.35595619
#>  [3,] 0.08567431 0.08567431
#>  [4,] 1.36220467 1.36220467
#>  [5,] 1.27878700 1.27878700
#>  [6,] 0.15237484 0.15237484
#>  [7,] 0.60255025 0.60255025
#>  [8,] 0.09360955 0.09360955
#>  [9,] 0.38145170 0.38145170
#> [10,] 0.26831469 0.26831469