Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 2.35219784 2.35219784 2.93024911
#>  [2,] 7.05005327 0.59020656 0.18787169
#>  [3,] 0.55455927 0.55455927 0.55455927
#>  [4,] 3.26957449 0.06214378 0.06214378
#>  [5,] 0.37246339 0.02902574 0.02902574
#>  [6,] 0.96269416 0.96269416 0.96269416
#>  [7,] 0.02509674 0.02509674 0.02509674
#>  [8,] 0.55157139 0.75695774 0.76173379
#>  [9,] 2.18413357 0.81467213 0.84544482
#> [10,] 0.94511200 0.75925187 1.77321853
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07074145 0.66502850 1.30273866
#>  [2,] 0.58133547 0.03112670 0.07908102
#>  [3,] 0.02248829 0.72152728 1.17401030
#>  [4,] 0.16404383 1.64863310 2.12460393
#>  [5,] 0.91737409 1.24824110 1.31731371
#>  [6,] 1.40608751 0.05333586 1.86510692
#>  [7,] 0.51108081 2.78549991 3.56645415
#>  [8,] 0.22988985 0.82672344 0.79484231
#>  [9,] 0.86484492 0.27313924 0.34951347
#> [10,] 1.68258708 0.12840958 2.87842156
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7030036 0.7030036 0.7030036
#>  [2,] 0.1324782 0.1324782 0.1324782
#>  [3,] 0.5259649 0.5259649 0.5259649
#>  [4,] 0.6164306 0.6164306 0.6164306
#>  [5,] 0.9399450 0.9399450 0.9399450
#>  [6,] 0.4560544 0.4560544 0.4560544
#>  [7,] 2.3569827 2.3569827 2.3569827
#>  [8,] 0.1870500 0.1870500 0.1870500
#>  [9,] 0.6075708 0.6075708 0.6075708
#> [10,] 1.5923066 1.5923066 1.5923066

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.52383974 1.2697024 1.26970236
#>  [2,] 0.46771430 0.4677143 0.46771430
#>  [3,] 2.31864624 2.6758608 0.80442720
#>  [4,] 0.18844798 0.7826816 1.00083825
#>  [5,] 0.23568058 0.2356806 0.34449606
#>  [6,] 2.07057946 1.6573659 0.05015472
#>  [7,] 2.11612185 2.1161219 2.11612185
#>  [8,] 0.09794229 1.3717334 0.09794229
#>  [9,] 0.64489703 0.6448970 0.69384507
#> [10,] 0.15282867 0.7815540 0.37939522
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.06554197 2.84519224 0.1238033
#>  [2,] 3.18549139 1.50024402 0.7764391
#>  [3,] 0.34841772 0.11704722 0.5733201
#>  [4,] 0.54228776 2.07213980 1.2651295
#>  [5,] 0.99813361 0.09936566 0.5441441
#>  [6,] 0.76035609 0.94188658 0.3309585
#>  [7,] 0.38498018 0.69121610 0.7272456
#>  [8,] 0.12055515 0.65252349 0.5627005
#>  [9,] 0.04020201 1.19524152 1.4263901
#> [10,] 1.85155390 1.57445683 0.7615428
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 5.4297821 5.4297821 5.4297821
#>  [2,] 0.8467285 0.8467285 0.8467285
#>  [3,] 0.4869421 0.4869421 0.4869421
#>  [4,] 1.4254190 1.4254190 1.4254190
#>  [5,] 0.2195769 0.2195769 0.2195769
#>  [6,] 0.2250821 0.2250821 0.2250821
#>  [7,] 0.6649344 0.6649344 0.6649344
#>  [8,] 1.8207457 1.8207457 1.8207457
#>  [9,] 0.4396890 0.4396890 0.4396890
#> [10,] 1.2240088 1.2240088 1.2240088

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02763761 0.02763761 0.02763761
#>  [2,] 1.47053380 1.47053380 1.47053380
#>  [3,] 0.52282822 0.52282822 0.27219489
#>  [4,] 2.26512164 2.26512164 1.84837366
#>  [5,] 0.25977595 0.25977595 0.25977595
#>  [6,] 2.44064385 2.72680968 2.72680968
#>  [7,] 1.87192709 1.80202212 1.32497833
#>  [8,] 0.33297173 0.39722596 1.10444133
#>  [9,] 3.61504571 2.54379118 0.59941778
#> [10,] 0.42051714 0.41199277 4.52033159
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>              [,1]        [,2]       [,3]
#>  [1,] 0.528017191 0.553344937 1.41106805
#>  [2,] 0.720734883 0.196535135 1.65254206
#>  [3,] 0.186763264 0.841279426 2.54381541
#>  [4,] 0.923228286 0.157060970 1.38491895
#>  [5,] 0.003318769 0.939284244 0.31797206
#>  [6,] 0.290410952 0.003560944 1.30665396
#>  [7,] 1.131220181 0.846203244 1.54319896
#>  [8,] 0.829370313 0.118219642 0.03174021
#>  [9,] 2.552203467 1.303320575 0.35220787
#> [10,] 0.139363010 0.101531537 1.90961546
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34046924 0.34046924 0.34046924
#>  [2,] 0.93699931 0.93699931 0.93699931
#>  [3,] 0.98899456 0.98899456 0.98899456
#>  [4,] 1.39760619 1.39760619 1.39760619
#>  [5,] 0.18516404 0.18516404 0.18516404
#>  [6,] 0.07696581 0.07696581 0.07696581
#>  [7,] 0.51134524 0.51134524 0.51134524
#>  [8,] 0.35388136 0.35388136 0.35388136
#>  [9,] 0.86087895 0.86087895 0.86087895
#> [10,] 0.67861748 0.67861748 0.67861748