Sample with Lévy-frailty model with compound Poisson subordinator

Draws n independent samples from a d-variate extendible Marshall-Olkin distribution corresponding to a LFM with a compound Poisson subordinator.

rextmo_lfm(
  n,
  d,
  rate,
  rate_killing,
  rate_drift,
  rjump_name,
  rjump_arg_list = list()
)

Arguments

n
Number of samples

d
Dimension

rate
Jump intensity of CPP subordinator

rate_killing
Killing intensity of CPP subordinator

rate_drift
Drift of CPP subordinator

rjump_name
Name of jump sampling function for jumps of CPP subordinator

rjump_arg_list
A list with named arguments for jump sampling function for jumps of CPP subordinator

Value

rextmo_lfm implements the Lévy-frailty model algorithm with a compound Poisson subordinator and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of the corresponding d-variate extendible Marshall-Olkin distribution.

Details

Model: The default times are defined as first exit times of a (killed) compound Poisson subordinator \(\Lambda\) with killing intensity \(a\), drift \(b\), jump intensity \(\xi\), and jump distribution \(X\).

Model parameters:

rate is the jump intensity of the compound Poisson subordinator.

rate_killing is the killing intensity of the compound Poisson subordinator.

rate_drift is the drift of the compound Poisson subordinator.

rjump_name and rjump_arg_list are the jump distribution name and list with jump distribution parameters. Currently, "rexp", "rposval", and "rpareto" are possible.

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 140 psqq.

See also

Examples

rextmo_lfm(10L, 2L, 0.5, 0.1, 0.2, "rposval", list("value"=1))

#>             [,1]       [,2]
#>  [1,] 0.53360772 0.53360772
#>  [2,] 0.13953948 0.67100408
#>  [3,] 0.11122485 0.33954229
#>  [4,] 1.03663352 1.02499172
#>  [5,] 2.98178618 2.62819470
#>  [6,] 2.04007718 0.62798332
#>  [7,] 0.46765823 0.86155821
#>  [8,] 1.60280850 0.69379130
#>  [9,] 0.85674310 0.85674310
#> [10,] 0.06959359 0.06959359

rextmo_lfm(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))

#>             [,1]       [,2]
#>  [1,]  1.2051005 14.5363947
#>  [2,]  1.6691409  1.6691409
#>  [3,]  7.7999790  5.1678548
#>  [4,] 15.7125469  5.3450053
#>  [5,]  0.2362183  0.2362183
#>  [6,]  1.4636057  1.4636057
#>  [7,]  2.7492549  4.2155902
#>  [8,]  0.3736335  7.2411320
#>  [9,]  7.2778702  6.0495844
#> [10,]  0.2116008  0.2116008

rextmo_lfm(10L, 2L, 0.5, 0, 0, "rpareto", list("alpha"=log2(2 - 0.5), x0 = 1e-4))

#>            [,1]      [,2]
#>  [1,]  53.98252 207.73648
#>  [2,] 311.05216 193.51886
#>  [3,]  67.77939 135.28008
#>  [4,]  14.45430  82.35803
#>  [5,] 115.11769 322.60920
#>  [6,] 240.75283 240.75283
#>  [7,] 163.88851 163.88851
#>  [8,] 134.22049 364.71768
#>  [9,] 420.42407 771.21685
#> [10,] 213.45584  35.33269

rextmo_lfm(10L, 2L, 0, 0, 1, "rposval", list("value"=1))  ## independence

#>             [,1]        [,2]
#>  [1,] 0.45099181 0.259793823
#>  [2,] 0.48141067 0.931376344
#>  [3,] 0.65076527 1.766176285
#>  [4,] 0.08050842 0.186904407
#>  [5,] 0.66815137 0.703072602
#>  [6,] 0.40455994 2.648782296
#>  [7,] 0.22342756 0.589334122
#>  [8,] 0.86847268 0.001655062
#>  [9,] 0.65647407 0.392622327
#> [10,] 0.16017312 0.108483846

rextmo_lfm(10L, 2L, 0, 1, 0, "rposval", list("value"=1))  ## comonotone

#>             [,1]       [,2]
#>  [1,] 0.52666119 0.52666119
#>  [2,] 0.07336072 0.07336072
#>  [3,] 2.19648856 2.19648856
#>  [4,] 0.69430780 0.69430780
#>  [5,] 0.50395103 0.50395103
#>  [6,] 1.01516116 1.01516116
#>  [7,] 0.50437277 0.50437277
#>  [8,] 0.68533094 0.68533094
#>  [9,] 2.74712209 2.74712209
#> [10,] 2.50375163 2.50375163