Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6506547 1.23877572 3.4979391
#>  [2,] 0.9372179 0.96963765 0.2063014
#>  [3,] 0.3054483 0.22955286 0.2295529
#>  [4,] 1.3091732 1.30917317 1.1281670
#>  [5,] 1.1354681 1.13546815 0.1639363
#>  [6,] 0.4182209 0.24833015 0.4182209
#>  [7,] 1.0560478 1.45967935 1.6505650
#>  [8,] 0.1006973 0.42951995 0.3761255
#>  [9,] 1.4659517 0.74014564 2.2753308
#> [10,] 5.9020023 0.06342227 0.7132389
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1580243 0.68369214 0.3246103
#>  [2,] 0.2727188 0.82273275 0.9079534
#>  [3,] 2.1472636 0.10036722 0.2549111
#>  [4,] 0.1018979 4.42525755 0.1345506
#>  [5,] 0.0542802 1.14924522 2.0280040
#>  [6,] 0.1251597 1.21403417 0.4930962
#>  [7,] 2.5950346 0.08664554 1.6358797
#>  [8,] 3.4398145 1.87688564 0.6115020
#>  [9,] 0.4786487 0.65682531 2.0187323
#> [10,] 0.1025913 0.40744520 0.2984303
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.63899659 0.63899659 0.63899659
#>  [2,] 0.58769123 0.58769123 0.58769123
#>  [3,] 1.03604987 1.03604987 1.03604987
#>  [4,] 0.83071786 0.83071786 0.83071786
#>  [5,] 0.35706242 0.35706242 0.35706242
#>  [6,] 1.47743252 1.47743252 1.47743252
#>  [7,] 0.06654196 0.06654196 0.06654196
#>  [8,] 0.88414040 0.88414040 0.88414040
#>  [9,] 0.87241041 0.87241041 0.87241041
#> [10,] 2.41510866 2.41510866 2.41510866

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.19233812 1.19233812 1.49619382
#>  [2,] 1.01230910 1.01230910 1.01230910
#>  [3,] 1.00457308 1.43732362 1.00457308
#>  [4,] 0.45281286 0.41692559 0.88829018
#>  [5,] 1.15150069 1.15150069 0.23110776
#>  [6,] 2.51125168 0.84576424 0.84265485
#>  [7,] 2.08705904 3.42414590 0.81739814
#>  [8,] 0.38673260 0.01238541 0.01238541
#>  [9,] 1.28905811 1.28905811 2.25744980
#> [10,] 0.05710887 0.05710887 0.05710887
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.62228205 2.0117029 0.8596494
#>  [2,] 0.48885239 0.9427497 0.8508861
#>  [3,] 1.90024497 0.3416342 0.7190542
#>  [4,] 0.81120020 1.7349188 1.3703173
#>  [5,] 0.05960605 0.3876806 2.4188365
#>  [6,] 3.82044229 1.4591778 2.2246340
#>  [7,] 0.48524731 0.3934644 0.2938994
#>  [8,] 0.38130572 0.8808037 0.2672316
#>  [9,] 2.90822472 0.5113734 0.7991007
#> [10,] 0.67458886 0.1513850 1.3637896
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.50726072 2.50726072 2.50726072
#>  [2,] 0.71360332 0.71360332 0.71360332
#>  [3,] 0.46012322 0.46012322 0.46012322
#>  [4,] 3.33283386 3.33283386 3.33283386
#>  [5,] 0.09029915 0.09029915 0.09029915
#>  [6,] 2.79981640 2.79981640 2.79981640
#>  [7,] 0.88611853 0.88611853 0.88611853
#>  [8,] 1.24212240 1.24212240 1.24212240
#>  [9,] 0.53127178 0.53127178 0.53127178
#> [10,] 0.65813737 0.65813737 0.65813737

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 2.0189626 2.93885850 2.7241373
#>  [2,] 0.1324618 0.13246182 0.1324618
#>  [3,] 0.2474993 1.10463533 1.6867266
#>  [4,] 1.0848546 1.99963664 1.9996366
#>  [5,] 1.6216972 0.17477492 5.2934502
#>  [6,] 0.6522120 1.42709549 1.4270955
#>  [7,] 0.9065582 0.29719348 0.9065582
#>  [8,] 0.7330609 2.82691400 0.5941186
#>  [9,] 0.2264502 0.08978817 1.2099531
#> [10,] 4.7335132 1.47626361 2.0154850
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.89209373 0.7562070 0.50906530
#>  [2,] 2.21666640 0.5356681 0.29502336
#>  [3,] 0.05399623 1.6776192 0.75653613
#>  [4,] 0.66557607 1.2306989 0.06206026
#>  [5,] 0.27061244 0.2030571 0.79080807
#>  [6,] 1.22007531 1.2314399 1.27373760
#>  [7,] 0.77175428 1.2773742 2.07349115
#>  [8,] 3.17995166 1.8499342 2.01443618
#>  [9,] 1.66902894 0.8767984 2.59219102
#> [10,] 0.25315818 1.1661100 1.04058511
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47089477 0.47089477 0.47089477
#>  [2,] 1.04030962 1.04030962 1.04030962
#>  [3,] 5.13519760 5.13519760 5.13519760
#>  [4,] 0.61730276 0.61730276 0.61730276
#>  [5,] 0.47597944 0.47597944 0.47597944
#>  [6,] 2.34241181 2.34241181 2.34241181
#>  [7,] 0.59751666 0.59751666 0.59751666
#>  [8,] 0.12094924 0.12094924 0.12094924
#>  [9,] 1.11146707 1.11146707 1.11146707
#> [10,] 0.04405025 0.04405025 0.04405025