Sample from a Marshall--Olkin distribution

Draws n independent samples from a d-variate Marshall-Olkin distribution with shock rates intensities.

rmo_esm(n, d, intensities)

rmo_arnold(n, d, intensities)

Arguments

n
Number of samples

d
Dimension

intensities
Marshall-Olkin intensity rates

Value

rmo_esm implements the exogenous shock model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

rmo_arnold implements the Arnold model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

Details

The shock intensities:

The shock intensities must be stored in a vector of length \(2^d-1\).

A shock intensity of zero corresponds to an almost surely infinite shock.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\) corresponding to the integer \(j\) if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

The exogenous shock model simulates a Marshall--Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

The Arnold model simulates a Marshall--Olkin distributed random variable by simulating a marked homogeneous Poisson process and where the inter-arrival times correspond to shock shock-arrival times and the marks to the specific shocks, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 , https://doi.org/10.1142/10265.

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 , https://doi.org/10.2307/2282907.

See also

Examples

rmo_esm(10L, 2L, c(0.4, 0.3, 0.2))

#>             [,1]      [,2]
#>  [1,] 0.15890160 0.1589016
#>  [2,] 0.66869058 0.6686906
#>  [3,] 0.02299967 4.1977187
#>  [4,] 0.73390652 1.0214477
#>  [5,] 0.51290672 0.5129067
#>  [6,] 3.58416305 0.2687313
#>  [7,] 0.35292350 1.1400560
#>  [8,] 1.53699397 1.5369940
#>  [9,] 0.04613704 0.1753271
#> [10,] 4.61329429 7.7756460

rmo_esm(10L, 2L, c(1, 1, 0))         ## independence

#>             [,1]       [,2]
#>  [1,] 0.45635995 0.04032429
#>  [2,] 1.10544443 0.48603647
#>  [3,] 0.04856027 0.70334948
#>  [4,] 0.24956238 2.27357966
#>  [5,] 0.37161926 0.52211881
#>  [6,] 0.12877145 0.55690861
#>  [7,] 0.27812198 1.20899012
#>  [8,] 1.13126986 0.68846283
#>  [9,] 0.81724429 0.42682390
#> [10,] 4.05844190 0.25291299

rmo_esm(10L, 2L, c(0, 0, 1))         ## comonotone

#>            [,1]      [,2]
#>  [1,] 0.6558426 0.6558426
#>  [2,] 0.2861381 0.2861381
#>  [3,] 0.4169685 0.4169685
#>  [4,] 0.1200703 0.1200703
#>  [5,] 2.4857798 2.4857798
#>  [6,] 0.2430092 0.2430092
#>  [7,] 4.2671167 4.2671167
#>  [8,] 0.9789015 0.9789015
#>  [9,] 0.1017248 0.1017248
#> [10,] 1.6371789 1.6371789

rmo_arnold(10L, 2L, c(0.4, 0.3, 0.2))

#>             [,1]      [,2]
#>  [1,] 3.49847789 6.6233033
#>  [2,] 1.21325454 2.6894353
#>  [3,] 1.70529796 2.3511802
#>  [4,] 0.17473521 1.3645833
#>  [5,] 0.08706506 8.2271991
#>  [6,] 4.20178030 0.8243962
#>  [7,] 1.75108571 3.8454167
#>  [8,] 0.59780783 0.6836592
#>  [9,] 1.85406192 3.0872776
#> [10,] 2.13640189 1.6033519

rmo_arnold(10L, 2L, c(1, 1, 0))         ## independence

#>             [,1]        [,2]
#>  [1,] 0.03819245 2.817854341
#>  [2,] 0.75867044 0.449562066
#>  [3,] 0.35197290 1.065217555
#>  [4,] 0.04113247 3.710015428
#>  [5,] 0.07095734 1.989546802
#>  [6,] 0.31650935 0.157298437
#>  [7,] 0.57509845 1.341311387
#>  [8,] 0.97490352 0.008309862
#>  [9,] 2.92105275 0.117876347
#> [10,] 2.65596120 0.140455950

rmo_arnold(10L, 2L, c(0, 0, 1))         ## comonotone

#>            [,1]      [,2]
#>  [1,] 0.1394811 0.1394811
#>  [2,] 3.0506897 3.0506897
#>  [3,] 0.1331726 0.1331726
#>  [4,] 0.8915171 0.8915171
#>  [5,] 1.6604923 1.6604923
#>  [6,] 1.4064678 1.4064678
#>  [7,] 0.6185939 0.6185939
#>  [8,] 2.5829326 2.5829326
#>  [9,] 0.4287433 0.4287433
#> [10,] 1.6027846 1.6027846