Sample from Cuadras-Auge distribution

Draws n independent samples from a d variate Cuadras-Augé distribution with parameters alpha and beta.

rmo_esm_cuadras_auge(n, d, alpha, beta)

Arguments

n
number of samples

d
dimension

alpha
rate of individual shocks

beta
rate of global shock

Value

rmo_esm_cuadras_auge implements an optimized version of the exogenous shock model representation for the Cuadras-Augé family and returns an \(n \times d\) array matrix with rows corresponding to the independent samples of size \(d\).

Details

alpha is the shock intensity of shocks that affect only single components.

beta is the shock intensity of the global shock that affects all components.

See also

Examples

rmo_esm_cuadras_auge(10L, 2L, 0.5, 0.2)

#>             [,1]       [,2]
#>  [1,] 1.84448426 0.16123876
#>  [2,] 0.28233880 1.44215771
#>  [3,] 1.14005604 1.14005604
#>  [4,] 0.03690963 0.10519624
#>  [5,] 3.69063543 4.37326292
#>  [6,] 0.91271989 0.08064858
#>  [7,] 0.97207295 0.09712054
#>  [8,] 0.49912475 3.51674739
#>  [9,] 1.04423762 0.25754291
#> [10,] 0.55624395 2.41798025

rmo_esm_cuadras_auge(10L, 2L, 0, 1)      ## comonotone

#>            [,1]      [,2]
#>  [1,] 1.1312699 1.1312699
#>  [2,] 0.6884628 0.6884628
#>  [3,] 0.8172443 0.8172443
#>  [4,] 0.4268239 0.4268239
#>  [5,] 4.0584419 4.0584419
#>  [6,] 0.2529130 0.2529130
#>  [7,] 0.6558426 0.6558426
#>  [8,] 0.2861381 0.2861381
#>  [9,] 0.4169685 0.4169685
#> [10,] 0.1200703 0.1200703

rmo_esm_cuadras_auge(10L, 2L, 1, 0)      ## independence

#>             [,1]       [,2]
#>  [1,] 2.48577980 0.24300920
#>  [2,] 4.26711665 0.97890153
#>  [3,] 0.10172479 1.63717892
#>  [4,] 3.14863010 0.36928823
#>  [5,] 0.02880499 0.04497353
#>  [6,] 0.52422183 2.33358958
#>  [7,] 1.45435566 0.61939240
#>  [8,] 0.11876648 0.46745937
#>  [9,] 1.09192908 0.25013858
#> [10,] 1.53476817 0.36159723