Sample from an exchangeable MO distribution

Draws n independent samples from a d variate exchangeable Marshall-Olkin distribution with shock rates ex_intensities.

rmo_ex_arnold(n, d, ex_intensities)

Arguments

n
number of samples

d
dimension

ex_intensities
exchangeable Marshall-Olkin intensity rates

Value

rmo_ex_arnold implements the modified Arnold model for the exchangeable subclass and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically disctributed samples of a \(d\) variate exchangeable Marshall-Olkin distribution with exchangeable parameters ex_intensities.

Details

The exchangeable shock intensities must be stored in a vector of length \(d\).

The entry \({exintensities}_{i}\) is the intensity of a shock corresponding to a set with \(i\) elements.

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 122 psqq.

See also

Examples

rmo_ex_arnold(10, 2, c(0.4, 0.2))

#>              [,1]       [,2]
#>  [1,] 4.550558496 4.55055850
#>  [2,] 1.791643456 0.88338730
#>  [3,] 1.827775187 0.50049577
#>  [4,] 2.812433416 1.97830060
#>  [5,] 0.254677182 0.03178032
#>  [6,] 1.748586625 1.25931561
#>  [7,] 0.002494141 1.93238744
#>  [8,] 5.160612172 1.43366522
#>  [9,] 1.733712602 0.14116940
#> [10,] 0.307398794 6.72997488

rmo_ex_arnold(10, 2, c(1, 0))      ## independence

#>             [,1]        [,2]
#>  [1,] 2.77002009 0.437326292
#>  [2,] 0.26575979 0.020162144
#>  [3,] 0.02428013 0.338337660
#>  [4,] 0.30014873 0.003146105
#>  [5,] 0.53918138 0.261059405
#>  [6,] 0.36148110 0.344231416
#>  [7,] 0.14306903 0.963466539
#>  [8,] 0.08622219 0.329231388
#>  [9,] 1.86211402 0.224935097
#> [10,] 0.01440250 0.700596381

rmo_ex_arnold(10, 2, c(0, 1))      ## comonotone

#>            [,1]      [,2]
#>  [1,] 2.3335896 2.3335896
#>  [2,] 0.1187665 0.1187665
#>  [3,] 0.2501386 0.2501386
#>  [4,] 0.6772600 0.6772600
#>  [5,] 0.1572617 0.1572617
#>  [6,] 0.4622073 0.4622073
#>  [7,] 2.5286583 2.5286583
#>  [8,] 0.2205992 0.2205992
#>  [9,] 0.2241656 0.2241656
#> [10,] 1.0513685 1.0513685