Sample with LFM and CPP subordinator

Draws n independent samples from a d-variate extendible Marshall-Olkin distribution corresponding to a LFM with a compound Poisson subordinator with parameters rate, rate_killing, rate_drift, rjump_name, and rjump_arg_list.

rmo_lfm_cpp(
  n,
  d,
  rate,
  rate_killing,
  rate_drift,
  rjump_name,
  rjump_arg_list = list()
)

Arguments

n
number of samples

d
dimension

rate
rate of CPP subordinator

rate_killing
killing rate of CPP subordinator

rate_drift
drift rate of CPP subordinator

rjump_name
name of jump sampling function for jumps of CPP subordinator

rjump_arg_list
list with named arguments for jump sampling function for jumps of CPP subordinator

Value

rmo_lfm_cpp implements the Lévy-frailty model representation with a compound Poisson subordinator and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of the corresponding d-variate extendible Marshall-Olkin distribution.

Details

rate is the jump intensity of the compound Poisson subordinator: at each given point-in-time, the waiting time to the next jump is exponentially distributed with rate rate.

rate_killing is the killing intensity of the compound Poisson subordinator: the probability that the compound Poisson subordinator jumps to its graveyard-state \(\infty\) between \(t\) and \(s\) is \(1 - \exp{\{-{ratekilling} (s-t)\}}\).

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 140 psqq.

See also

Examples

rmo_lfm_cpp(10L, 2L, 0.5, 0.1, 0.2, "rposval", list("value"=1))

#>            [,1]      [,2]
#>  [1,] 0.4860184 3.8604655
#>  [2,] 4.8945077 0.5086240
#>  [3,] 0.4497353 0.1440250
#>  [4,] 1.6679479 1.1724517
#>  [5,] 1.1625880 2.6738408
#>  [6,] 0.2344509 0.1567171
#>  [7,] 4.3582362 1.1029962
#>  [8,] 0.5526044 0.5526044
#>  [9,] 0.9693047 0.7878149
#> [10,] 0.4922071 0.7329467

rmo_lfm_cpp(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))

#>             [,1]       [,2]
#>  [1,]  7.6766682  7.6766682
#>  [2,]  0.7585661  0.7585661
#>  [3,]  3.6423305  3.6423305
#>  [4,]  1.6269882  4.6085298
#>  [5,]  6.1571251  6.4251133
#>  [6,]  8.0175440  2.3281082
#>  [7,] 10.4896824 13.3182855
#>  [8,] 17.2426682 18.9168472
#>  [9,]  3.3909483  1.2996616
#> [10,]  4.1571849 11.1076425

rmo_lfm_cpp(10L, 2L, 0, 0, 1, "rposval", list("value"=1))  ## independence

#>            [,1]      [,2]
#>  [1,] 0.2809119 0.7719257
#>  [2,] 0.6948640 1.5606747
#>  [3,] 0.2569003 0.9576092
#>  [4,] 0.0637486 0.4755903
#>  [5,] 0.1394811 0.3929772
#>  [6,] 0.4015350 3.0506897
#>  [7,] 1.0858749 0.1764943
#>  [8,] 0.4985932 0.8915171
#>  [9,] 0.2329973 1.6604923
#> [10,] 1.2995392 6.9243290

rmo_lfm_cpp(10L, 2L, 0, 1, 0, "rposval", list("value"=1))  ## comonotone

#>             [,1]       [,2]
#>  [1,] 0.58895533 0.58895533
#>  [2,] 1.25410909 1.25410909
#>  [3,] 1.96600090 1.96600090
#>  [4,] 0.38183566 0.38183566
#>  [5,] 0.95607280 0.95607280
#>  [6,] 0.02670447 0.02670447
#>  [7,] 2.13003743 2.13003743
#>  [8,] 0.38143820 0.38143820
#>  [9,] 2.66359687 2.66359687
#> [10,] 1.27157350 1.27157350