Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3959611 2.0137667 0.1395196
#>  [2,] 0.5699935 0.6578634 1.2519304
#>  [3,] 0.1826477 1.2222726 0.1966954
#>  [4,] 2.0601570 1.9055233 2.0601570
#>  [5,] 1.3912104 0.9669611 3.1338655
#>  [6,] 0.0472192 5.4131457 0.5411522
#>  [7,] 2.1869968 0.5010856 0.1433743
#>  [8,] 3.1924305 2.0235201 1.3746702
#>  [9,] 1.3849146 2.0454363 2.8180502
#> [10,] 1.6341303 0.2294068 2.2777615
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.6904954 2.6904954 2.6904954
#>  [2,] 0.8422141 0.8422141 0.8422141
#>  [3,] 1.2576537 1.2576537 1.2576537
#>  [4,] 0.4193473 0.4193473 0.4193473
#>  [5,] 0.4416840 0.4416840 0.4416840
#>  [6,] 1.2212989 1.2212989 1.2212989
#>  [7,] 0.8395134 0.8395134 0.8395134
#>  [8,] 0.5463166 0.5463166 0.5463166
#>  [9,] 0.4138898 0.4138898 0.4138898
#> [10,] 2.8009082 2.8009082 2.8009082
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43809071 0.04452634 3.55157719
#>  [2,] 0.54304271 1.32526625 0.07243627
#>  [3,] 0.23637550 2.41934377 2.12597342
#>  [4,] 2.15080677 0.33175384 0.10739776
#>  [5,] 0.01775727 1.47338934 0.70223791
#>  [6,] 2.98649385 2.02868902 1.62211920
#>  [7,] 2.78862694 0.20184927 0.57009032
#>  [8,] 0.17566345 0.59221933 1.90322248
#>  [9,] 0.55113937 0.88684549 0.95021025
#> [10,] 5.34387593 0.11679385 2.78744594

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.05056725 7.08667164 0.3078102
#>  [2,] 1.25793463 0.52164405 4.7936063
#>  [3,] 4.56773917 0.58691023 0.1360696
#>  [4,] 1.44419988 1.44419988 1.2355511
#>  [5,] 1.13723042 0.05126607 0.3248015
#>  [6,] 0.03603507 0.10758338 0.8934298
#>  [7,] 1.72676118 3.53705515 0.9683390
#>  [8,] 0.65316300 1.98120193 0.8656972
#>  [9,] 5.25456005 0.22581902 2.0205191
#> [10,] 0.60808768 0.60808768 0.6080877
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.0671509 3.0671509 3.0671509
#>  [2,] 0.4688110 0.4688110 0.4688110
#>  [3,] 0.7326971 0.7326971 0.7326971
#>  [4,] 0.4195476 0.4195476 0.4195476
#>  [5,] 0.2535079 0.2535079 0.2535079
#>  [6,] 0.5589991 0.5589991 0.5589991
#>  [7,] 2.0764236 2.0764236 2.0764236
#>  [8,] 0.5206076 0.5206076 0.5206076
#>  [9,] 0.3699501 0.3699501 0.3699501
#> [10,] 0.6978047 0.6978047 0.6978047
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.02929292 0.08848288 0.7929109
#>  [2,] 0.45056160 1.31036537 0.3552411
#>  [3,] 0.42210270 1.70906816 0.1366564
#>  [4,] 0.13838423 1.13443422 0.8131170
#>  [5,] 0.66875205 1.91181871 4.3800438
#>  [6,] 0.11296945 2.12769568 0.3087545
#>  [7,] 0.46553503 0.23348300 1.5876680
#>  [8,] 1.05400403 0.78681467 0.4620363
#>  [9,] 0.49656761 1.42074864 0.9306866
#> [10,] 0.60645160 0.49865131 0.8904794

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.3724023 4.3487240 4.59964713
#>  [2,] 0.1102460 4.9917747 4.42316748
#>  [3,] 0.1550315 3.1852090 1.21685990
#>  [4,] 0.8387581 0.8387581 0.83875807
#>  [5,] 1.4048829 1.0495552 0.92330380
#>  [6,] 1.1230378 0.4017653 1.71206944
#>  [7,] 0.3381439 0.3381439 0.06490124
#>  [8,] 0.9431209 3.8127266 0.40833002
#>  [9,] 1.6641499 0.9109472 1.55531450
#> [10,] 1.8308829 2.8007424 2.51615368
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.1359286 2.1359286 2.1359286
#>  [2,] 0.7401976 0.7401976 0.7401976
#>  [3,] 0.1481921 0.1481921 0.1481921
#>  [4,] 0.2152662 0.2152662 0.2152662
#>  [5,] 0.7793816 0.7793816 0.7793816
#>  [6,] 1.2019367 1.2019367 1.2019367
#>  [7,] 3.4305611 3.4305611 3.4305611
#>  [8,] 0.2499093 0.2499093 0.2499093
#>  [9,] 0.2804355 0.2804355 0.2804355
#> [10,] 0.1304931 0.1304931 0.1304931
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5888091 0.02935969 2.8987671
#>  [2,] 0.5193942 1.80916520 0.1276181
#>  [3,] 0.1055248 0.81133161 1.1832383
#>  [4,] 0.1542669 0.96700326 0.4251091
#>  [5,] 1.8016174 0.02842022 1.1768204
#>  [6,] 0.1502981 3.70869408 0.1952395
#>  [7,] 0.4535044 0.15477159 0.1200415
#>  [8,] 0.9830892 0.12066048 2.7439013
#>  [9,] 1.2980496 0.71797848 1.1612920
#> [10,] 0.2477112 1.85579562 1.6939854

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.787636261 0.669500222 1.015641794
#>  [2,] 0.110104963 0.995003656 1.379366164
#>  [3,] 3.352184431 1.130511638 0.810690278
#>  [4,] 0.008317704 0.008317704 0.008317704
#>  [5,] 1.302683991 2.154916708 0.694694388
#>  [6,] 0.588791076 0.588791076 0.074246313
#>  [7,] 0.610596214 1.257130691 1.494613158
#>  [8,] 0.336922723 3.092800966 0.507546341
#>  [9,] 0.805612366 1.247363843 1.350141917
#> [10,] 6.292123387 2.292550946 2.816261941
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6511174 0.6511174 0.6511174
#>  [2,] 0.2550141 0.2550141 0.2550141
#>  [3,] 0.2706211 0.2706211 0.2706211
#>  [4,] 0.1993871 0.1993871 0.1993871
#>  [5,] 0.1991642 0.1991642 0.1991642
#>  [6,] 0.3529390 0.3529390 0.3529390
#>  [7,] 0.6009312 0.6009312 0.6009312
#>  [8,] 1.0421968 1.0421968 1.0421968
#>  [9,] 0.5358767 0.5358767 0.5358767
#> [10,] 2.7328840 2.7328840 2.7328840
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29139640 2.01963978 0.95592499
#>  [2,] 0.62069928 0.24193868 5.14641086
#>  [3,] 0.55911800 1.77975935 0.69075688
#>  [4,] 0.59620014 1.29450172 0.67613094
#>  [5,] 0.86974276 0.23479341 0.51565810
#>  [6,] 0.82439076 0.72257906 0.26313538
#>  [7,] 1.51841011 1.19379983 0.17855354
#>  [8,] 2.21993197 0.07052564 0.65482781
#>  [9,] 0.13011928 3.02865572 0.78702282
#> [10,] 0.02890744 0.76755906 0.04420112

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.47557153 1.4826595 1.38131872
#>  [2,] 4.14960966 0.2545787 0.19895513
#>  [3,] 0.90700973 1.2715976 1.76991108
#>  [4,] 5.08733620 3.2825370 1.30672444
#>  [5,] 0.50049932 0.5004993 0.50049932
#>  [6,] 0.14668985 2.1757296 1.32120555
#>  [7,] 1.71407014 1.7140701 1.71407014
#>  [8,] 1.85128177 1.2861983 0.02747898
#>  [9,] 1.40077969 1.1510859 0.16698454
#> [10,] 0.09816034 0.8968528 1.23279448
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 3.12143812 3.12143812 3.12143812
#>  [2,] 0.07234275 0.07234275 0.07234275
#>  [3,] 0.16226909 0.16226909 0.16226909
#>  [4,] 0.34892590 0.34892590 0.34892590
#>  [5,] 1.13570028 1.13570028 1.13570028
#>  [6,] 0.37072188 0.37072188 0.37072188
#>  [7,] 0.68650678 0.68650678 0.68650678
#>  [8,] 0.73150762 0.73150762 0.73150762
#>  [9,] 1.26812277 1.26812277 1.26812277
#> [10,] 0.92831692 0.92831692 0.92831692
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2859633 0.7256424 0.1423267
#>  [2,] 0.7152257 0.1541817 0.2164858
#>  [3,] 0.3288287 1.1051641 1.1181122
#>  [4,] 0.2322474 3.4231898 1.0114539
#>  [5,] 0.3226531 0.1836117 0.7501281
#>  [6,] 0.3899566 1.5894596 0.9413069
#>  [7,] 0.1587404 2.1212611 0.5547300
#>  [8,] 0.8089894 1.4823606 1.0591873
#>  [9,] 0.1622915 1.2800591 2.4112196
#> [10,] 1.2944444 0.1943718 0.7842818

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0239742 0.05427883 1.74552660
#>  [2,] 0.6784197 0.67841969 0.67841969
#>  [3,] 0.6650372 3.78870976 3.64618293
#>  [4,] 0.7401571 0.13230810 1.30254668
#>  [5,] 1.8211453 0.52802589 2.16001808
#>  [6,] 2.0377306 0.27105156 0.88498738
#>  [7,] 1.8960186 2.10312795 2.10312795
#>  [8,] 3.2326582 5.65726542 0.09774574
#>  [9,] 2.8083539 4.48239302 0.96788450
#> [10,] 1.8552815 1.92784293 0.94070956
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33555899 0.33555899 0.33555899
#>  [2,] 0.71761823 0.71761823 0.71761823
#>  [3,] 1.66338485 1.66338485 1.66338485
#>  [4,] 0.53670335 0.53670335 0.53670335
#>  [5,] 0.63443429 0.63443429 0.63443429
#>  [6,] 2.84799882 2.84799882 2.84799882
#>  [7,] 1.08927029 1.08927029 1.08927029
#>  [8,] 0.44262871 0.44262871 0.44262871
#>  [9,] 0.08184413 0.08184413 0.08184413
#> [10,] 0.35974421 0.35974421 0.35974421
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.87744324 0.32395117 1.5179074
#>  [2,] 0.01757303 0.77930008 1.8350609
#>  [3,] 5.44907795 0.36601544 0.1331056
#>  [4,] 2.88901792 0.14979324 1.4805489
#>  [5,] 0.40101274 0.42987656 0.5968568
#>  [6,] 0.61095654 0.83993511 0.1334054
#>  [7,] 0.53645782 1.88803836 0.2659518
#>  [8,] 0.42617178 0.01112876 1.9275569
#>  [9,] 0.17751345 1.65784944 0.8096933
#> [10,] 2.18867409 1.48713823 2.4441315

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.03495364 0.52143958 0.5214396
#>  [2,] 0.47824896 1.39137146 0.3236935
#>  [3,] 4.89417755 0.08219852 0.6377889
#>  [4,] 0.73271315 0.80796278 0.1209135
#>  [5,] 0.41825649 3.00255981 2.4928666
#>  [6,] 2.16959957 0.32163191 1.2304943
#>  [7,] 0.18333909 1.34953868 1.9357739
#>  [8,] 1.91097633 0.64981415 1.9109763
#>  [9,] 2.33334714 1.78679076 2.3228917
#> [10,] 2.47161734 0.09616554 0.1189291
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32306819 0.32306819 0.32306819
#>  [2,] 1.19705553 1.19705553 1.19705553
#>  [3,] 0.60000934 0.60000934 0.60000934
#>  [4,] 0.26438437 0.26438437 0.26438437
#>  [5,] 2.58569530 2.58569530 2.58569530
#>  [6,] 0.92444341 0.92444341 0.92444341
#>  [7,] 0.05785914 0.05785914 0.05785914
#>  [8,] 1.39543682 1.39543682 1.39543682
#>  [9,] 0.29411470 0.29411470 0.29411470
#> [10,] 3.38038508 3.38038508 3.38038508
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.49667007 1.8732199 0.7472857
#>  [2,] 0.01082212 0.9249272 0.3483670
#>  [3,] 0.41613418 0.1654725 2.7057985
#>  [4,] 2.54892143 0.7244740 0.5350284
#>  [5,] 0.02048081 0.5305135 1.6009261
#>  [6,] 3.45183285 3.2646008 0.3855438
#>  [7,] 0.77447037 0.3036459 1.2187617
#>  [8,] 2.19910117 1.8964246 1.2916511
#>  [9,] 0.54820018 2.5394659 1.3590145
#> [10,] 1.26964297 1.1412262 0.6775273

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.18063330 2.4386958 0.8862321
#>  [2,] 2.17751761 2.1775176 1.4296679
#>  [3,] 0.45866016 0.9632183 0.9374341
#>  [4,] 0.37590131 0.3759013 0.3759013
#>  [5,] 1.49938724 1.1280342 0.4396121
#>  [6,] 0.63587099 0.6358710 0.6358710
#>  [7,] 0.35779271 1.0047031 2.3286326
#>  [8,] 0.04875915 1.1847445 3.1543770
#>  [9,] 2.68954712 2.6895471 0.9728825
#> [10,] 0.97798497 0.1006811 2.2964887
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.42768409 1.42768409 1.42768409
#>  [2,] 0.05130751 0.05130751 0.05130751
#>  [3,] 0.06813280 0.06813280 0.06813280
#>  [4,] 0.21714363 0.21714363 0.21714363
#>  [5,] 0.38851546 0.38851546 0.38851546
#>  [6,] 0.07088317 0.07088317 0.07088317
#>  [7,] 0.01363497 0.01363497 0.01363497
#>  [8,] 2.59762981 2.59762981 2.59762981
#>  [9,] 1.08947994 1.08947994 1.08947994
#> [10,] 0.44062231 0.44062231 0.44062231
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.03627727 1.62473832 1.5453712
#>  [2,] 1.25116048 0.03413681 0.4217563
#>  [3,] 1.04771694 0.61503037 1.3054652
#>  [4,] 0.46286124 1.07335219 1.2014989
#>  [5,] 0.63728862 0.53379247 0.3186711
#>  [6,] 1.83599150 0.18101335 0.4417075
#>  [7,] 1.04723641 0.65991591 0.7157017
#>  [8,] 0.78339276 0.62895147 0.6501404
#>  [9,] 1.26041871 2.11265956 0.4118970
#> [10,] 2.04099393 0.13633842 0.6764275

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 2.732898276 0.30875385 2.73289828
#>  [2,] 0.912697242 0.39770257 0.91269724
#>  [3,] 0.815995634 5.22693450 1.70296228
#>  [4,] 0.616064090 0.61606409 0.61606409
#>  [5,] 1.549409920 0.23745364 6.94847531
#>  [6,] 2.833832703 0.93743277 2.83383270
#>  [7,] 0.370452121 3.68013274 2.43412642
#>  [8,] 0.003652395 1.20893244 1.80996070
#>  [9,] 0.048974463 0.04897446 0.04897446
#> [10,] 3.049538451 1.51820872 0.52157128
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60103955 0.60103955 0.60103955
#>  [2,] 1.59139224 1.59139224 1.59139224
#>  [3,] 3.42924707 3.42924707 3.42924707
#>  [4,] 0.91924888 0.91924888 0.91924888
#>  [5,] 0.60172071 0.60172071 0.60172071
#>  [6,] 0.14741874 0.14741874 0.14741874
#>  [7,] 1.38856890 1.38856890 1.38856890
#>  [8,] 1.17922683 1.17922683 1.17922683
#>  [9,] 0.07858315 0.07858315 0.07858315
#> [10,] 2.18867504 2.18867504 2.18867504
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0870726 0.7002491 0.02027127
#>  [2,] 1.3091359 0.2123164 1.19663348
#>  [3,] 0.0282246 3.1933020 0.02411996
#>  [4,] 0.5647952 0.5324450 0.80782286
#>  [5,] 0.4332158 0.2277368 0.55696578
#>  [6,] 0.2839495 1.5948533 1.18676170
#>  [7,] 2.0343059 2.3795812 0.28393364
#>  [8,] 0.1449744 1.3600945 5.59078021
#>  [9,] 0.5870579 0.2603383 2.80342402
#> [10,] 0.2835606 0.4898702 0.01541361

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.19303437 0.1976588 1.0767948
#>  [2,] 1.75250944 1.1169071 2.4830084
#>  [3,] 0.99407314 0.9940731 0.6353208
#>  [4,] 0.59394488 0.5939449 0.5939449
#>  [5,] 3.48997404 1.5028342 0.5702998
#>  [6,] 0.19228602 4.6832973 5.3718710
#>  [7,] 0.07099652 1.1714443 2.9056526
#>  [8,] 1.88135663 2.5547576 1.9059905
#>  [9,] 2.36559623 0.5811455 2.3655962
#> [10,] 1.50150037 0.7531094 1.6070899
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07038468 0.07038468 0.07038468
#>  [2,] 0.74974674 0.74974674 0.74974674
#>  [3,] 0.70194087 0.70194087 0.70194087
#>  [4,] 1.43551044 1.43551044 1.43551044
#>  [5,] 3.79156370 3.79156370 3.79156370
#>  [6,] 0.79410647 0.79410647 0.79410647
#>  [7,] 0.14016886 0.14016886 0.14016886
#>  [8,] 0.09169360 0.09169360 0.09169360
#>  [9,] 0.69142641 0.69142641 0.69142641
#> [10,] 1.87804257 1.87804257 1.87804257
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.51561541 0.52950970 0.17674579
#>  [2,] 0.12650293 1.48855187 0.33352435
#>  [3,] 0.09634705 2.84281883 1.12694929
#>  [4,] 3.13240218 1.36591340 0.77861368
#>  [5,] 0.26625732 0.56999246 0.03845203
#>  [6,] 3.15177591 0.08773617 1.14717605
#>  [7,] 0.78975340 0.46110924 0.71191268
#>  [8,] 1.01004122 0.13584908 0.96968857
#>  [9,] 0.27395736 1.48723000 0.63259006
#> [10,] 0.82774088 0.33484944 0.40851892

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.92563440 0.3055182 0.23695558
#>  [2,] 0.90031285 1.2414777 1.03170074
#>  [3,] 0.30569567 0.2569911 0.66676538
#>  [4,] 0.09431081 1.2328437 0.09431081
#>  [5,] 0.07648060 0.1047377 0.10473771
#>  [6,] 1.47255039 0.3248377 2.47559578
#>  [7,] 0.39472898 0.3947290 0.39472898
#>  [8,] 0.23662590 2.7879692 0.52712409
#>  [9,] 0.17688275 0.3112979 1.27278080
#> [10,] 0.06978997 1.5146195 0.14066751
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35399209 0.35399209 1.08007700
#>  [2,] 0.04663925 0.92987613 0.04663925
#>  [3,] 0.29666843 0.29666843 0.29666843
#>  [4,] 0.06616127 0.02391638 1.13260169
#>  [5,] 1.21724768 0.06394628 0.06394628
#>  [6,] 0.58767720 0.04075819 0.08701699
#>  [7,] 0.59245015 0.50219695 0.11656587
#>  [8,] 0.04698937 0.32868075 0.04698937
#>  [9,] 0.24258116 1.41148439 0.57677648
#> [10,] 0.63688972 0.58618623 0.80127087
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 3.02525054 0.9939511 1.08859402
#>  [2,] 0.08675343 0.1717628 0.17176285
#>  [3,] 1.10214633 0.2373330 1.47234165
#>  [4,] 0.44450745 1.2532876 0.40217315
#>  [5,] 1.16624829 0.4500236 1.16624829
#>  [6,] 0.57881938 0.2926133 0.17994860
#>  [7,] 0.29754896 0.5158458 0.07371949
#>  [8,] 0.16102300 0.1610230 0.10725906
#>  [9,] 1.16432318 0.2984194 0.87390769
#> [10,] 0.46030287 0.4603029 1.36269728
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.2631382 0.38215347 0.38703550
#>  [2,] 0.8252983 0.73048144 0.08336471
#>  [3,] 4.8635092 0.28066291 1.14764672
#>  [4,] 0.5240093 0.64655296 0.16417736
#>  [5,] 0.5370748 0.53707477 0.53707477
#>  [6,] 0.8537378 0.02647437 0.24838510
#>  [7,] 0.5557075 0.39686801 0.15457725
#>  [8,] 0.9313975 0.57340558 1.43644636
#>  [9,] 2.1970260 0.56561492 0.60472006
#> [10,] 0.1364959 0.13649587 0.13649587
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.11807646 0.118076464 0.11807646
#>  [2,] 0.18114459 0.136939499 0.23333467
#>  [3,] 0.04220505 0.042205049 0.04220505
#>  [4,] 0.01504929 0.015049293 0.01504929
#>  [5,] 0.38815016 0.006461809 0.38815016
#>  [6,] 0.10497518 0.509976142 0.61691866
#>  [7,] 0.33097402 0.264874794 0.58451325
#>  [8,] 0.69315303 0.693153032 0.06100298
#>  [9,] 1.71617543 0.950867566 1.36268848
#> [10,] 0.23250140 0.733361391 0.19276120

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9257356 2.7515442 0.16444992
#>  [2,] 2.0951199 5.9831201 0.03074258
#>  [3,] 0.1349793 1.6777207 1.98626391
#>  [4,] 0.9642670 0.9382302 0.93139837
#>  [5,] 5.5209478 1.2714439 5.19685036
#>  [6,] 0.8139605 0.8923301 1.15985008
#>  [7,] 4.5980280 0.5094289 2.05230891
#>  [8,] 1.0911742 1.0911742 0.55748496
#>  [9,] 0.4360305 0.4360305 0.43603049
#> [10,] 1.6046309 1.2216224 2.52786090
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,]  9.23820629 4.14060738 0.09653356
#>  [2,]  0.32279744 0.32279744 0.32279744
#>  [3,]  0.30399060 0.83340235 0.15087657
#>  [4,] 11.13528944 1.03215866 0.75986342
#>  [5,]  1.90490031 3.18151707 0.76039884
#>  [6,]  2.34686821 0.68899338 2.34686821
#>  [7,]  1.14285701 2.41214800 1.97512901
#>  [8,]  2.52734015 0.04399524 0.86874464
#>  [9,]  0.80703961 0.36563446 0.36563446
#> [10,]  0.04175602 0.88682271 1.19788927
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.2587395 1.45141135 3.31049070
#>  [2,] 3.3370003 0.30561950 3.33700033
#>  [3,] 0.4848830 9.02805985 0.78792461
#>  [4,] 0.9768320 0.97073013 1.60002222
#>  [5,] 2.6913798 0.63183735 1.14442506
#>  [6,] 1.0547670 3.67039945 0.80282126
#>  [7,] 0.6554140 4.02482392 1.63475977
#>  [8,] 1.6427761 0.07011062 0.04480915
#>  [9,] 0.7276442 0.72764424 0.72764424
#> [10,] 0.6579420 1.43227009 1.43227009
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 1.030639311 1.030639311 0.261355316
#>  [2,] 1.073641195 2.306492539 0.079371795
#>  [3,] 0.226670181 0.858731277 0.881890787
#>  [4,] 0.562877747 0.687932449 2.743252320
#>  [5,] 0.917429468 0.538809822 0.217894424
#>  [6,] 1.379675526 1.290690733 0.170120621
#>  [7,] 2.344512054 0.975917002 0.243108939
#>  [8,] 0.963970221 0.208294999 1.509349188
#>  [9,] 2.796964524 0.031678222 0.274524259
#> [10,] 0.003030616 0.003030616 0.003030616
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36008772 0.20642876 0.36008772
#>  [2,] 0.09604893 3.58949120 0.41439572
#>  [3,] 0.06039349 0.02688006 0.06039349
#>  [4,] 0.38489807 1.57994548 0.43194918
#>  [5,] 0.05695716 0.67154227 0.67154227
#>  [6,] 0.78306475 5.89873207 0.84484208
#>  [7,] 1.72322447 0.50394115 0.03609455
#>  [8,] 3.11957584 1.97971658 0.69222752
#>  [9,] 0.68045170 0.46570364 4.94219656
#> [10,] 2.99618269 1.19302330 2.04288927

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24757594 0.54787218 0.36649357
#>  [2,] 0.45355232 0.45355232 0.45355232
#>  [3,] 0.55172010 0.55172010 0.55172010
#>  [4,] 0.07840269 0.07840269 0.07840269
#>  [5,] 0.47536872 0.83738029 0.59359813
#>  [6,] 1.25931883 1.25931883 0.76436530
#>  [7,] 0.17187113 0.17187113 0.17187113
#>  [8,] 1.39541068 1.16325109 1.42038766
#>  [9,] 0.52236775 0.09181547 0.52236775
#> [10,] 0.53571282 0.53571282 0.57570229
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.61112686 0.94553856 0.37741537
#>  [2,] 0.17990659 0.17990659 0.17990659
#>  [3,] 0.78650521 0.78650521 0.78650521
#>  [4,] 0.49761445 0.49761445 0.36554809
#>  [5,] 0.03418827 0.03418827 0.82119653
#>  [6,] 0.19160898 0.86202237 0.86202237
#>  [7,] 0.74345542 0.81624765 0.27432070
#>  [8,] 1.63695606 1.93102176 0.37043148
#>  [9,] 0.02843267 0.91836195 0.02843267
#> [10,] 0.96492333 0.96492333 0.02712038
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08897935 0.39621530 0.32992309
#>  [2,] 0.58751320 0.58751320 0.10426933
#>  [3,] 0.14986001 0.14986001 0.05189081
#>  [4,] 0.14546986 0.60070453 0.14546986
#>  [5,] 0.07311372 0.07311372 0.07311372
#>  [6,] 1.62116912 0.48428910 1.62116912
#>  [7,] 0.03954923 0.21759098 0.03954923
#>  [8,] 0.72370863 0.27767203 0.27767203
#>  [9,] 0.33741501 0.33741501 0.33741501
#> [10,] 0.44893672 0.44893672 0.44893672
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.673761687 0.67376169 0.58929699
#>  [2,] 1.131347574 2.00697183 2.35887107
#>  [3,] 0.003768167 0.31890362 0.31890362
#>  [4,] 0.240769436 0.24076944 0.24076944
#>  [5,] 0.855319830 0.47919175 0.85531983
#>  [6,] 0.679173927 1.06860426 0.02383266
#>  [7,] 0.467480245 0.46748025 0.02662293
#>  [8,] 0.044832914 0.02307601 0.04483291
#>  [9,] 0.256483934 0.25648393 0.25648393
#> [10,] 0.003329876 0.06838760 0.78448692
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1440628 0.23468243 0.23468243
#>  [2,] 0.1817088 0.40183453 1.57626422
#>  [3,] 0.3239697 3.23139965 0.10841377
#>  [4,] 1.0871813 1.42721982 1.22949222
#>  [5,] 0.4331171 0.56905017 0.67905780
#>  [6,] 0.4037507 0.27531913 0.40375073
#>  [7,] 0.5712096 0.00583519 0.14903160
#>  [8,] 0.3367801 0.90817824 0.33678011
#>  [9,] 0.3037648 0.30376483 0.30376483
#> [10,] 0.6384423 0.39806588 0.05857404

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07573917 1.27699714 0.20145970
#>  [2,] 0.01517954 0.61367884 0.05009166
#>  [3,] 0.01382268 0.01382268 0.01382268
#>  [4,] 0.21504399 0.21504399 0.17099766
#>  [5,] 1.50566968 1.06348468 0.68309890
#>  [6,] 0.04414083 0.33959316 0.04414083
#>  [7,] 0.17063419 0.17063419 0.07740664
#>  [8,] 0.01984601 0.01984601 0.01984601
#>  [9,] 1.31313337 0.02438431 0.15066906
#> [10,] 0.12481722 0.42236646 0.25188647
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3456831 0.10149076 0.3790280
#>  [2,] 0.5410357 0.01372078 0.5410357
#>  [3,] 0.3135741 0.31357409 0.3135741
#>  [4,] 0.1460076 0.14600760 0.1460076
#>  [5,] 0.3554824 0.28071111 0.3554824
#>  [6,] 1.2386540 1.01625989 1.1266419
#>  [7,] 0.3637318 0.36373179 0.2123978
#>  [8,] 0.2823457 1.74046229 1.7404623
#>  [9,] 0.7386849 0.70770353 0.7386849
#> [10,] 0.1817259 0.23675962 0.1493547
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05733993 0.05733993 0.05733993
#>  [2,] 0.73226168 0.25771203 0.20820443
#>  [3,] 0.38920372 0.38920372 0.38920372
#>  [4,] 0.29661767 0.29661767 0.09073928
#>  [5,] 0.20911805 0.20911805 0.20911805
#>  [6,] 0.68080678 0.12865464 0.67553211
#>  [7,] 0.22450550 0.15305801 0.22450550
#>  [8,] 0.16566923 0.04790050 0.04790050
#>  [9,] 1.37706391 1.37706391 1.43378878
#> [10,] 0.16084502 0.16084502 0.16084502
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19707416 0.19707416 0.19707416
#>  [2,] 0.61395718 0.10457161 0.07849635
#>  [3,] 0.93062565 0.13852829 0.03905879
#>  [4,] 0.75374643 0.75374643 0.09258838
#>  [5,] 0.31078827 0.79845511 0.79845511
#>  [6,] 0.16232065 0.01105271 0.01900579
#>  [7,] 0.54741709 0.18150642 0.41704549
#>  [8,] 0.09748899 0.58004108 1.05354558
#>  [9,] 0.23056409 0.14898907 0.18843946
#> [10,] 0.10619601 0.10619601 0.10619601

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>              [,1]      [,2]       [,3]
#>  [1,] 0.053568108 0.2649588 0.29870707
#>  [2,] 0.344815602 0.9054676 0.09186154
#>  [3,] 0.004963212 0.7483379 0.11928656
#>  [4,] 0.040507746 0.1589052 0.15890516
#>  [5,] 0.337196536 0.4876487 0.16707997
#>  [6,] 0.123995458 0.3007143 0.30071426
#>  [7,] 0.112022782 0.1120228 0.11202278
#>  [8,] 0.029210995 0.1443341 0.67033819
#>  [9,] 0.151235191 0.6443953 1.02022551
#> [10,] 0.649965578 0.5679693 0.70159858
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>             [,1]       [,2]         [,3]
#>  [1,] 1.08307898 0.64884295 0.1708260107
#>  [2,] 0.05070527 0.05070527 0.0802860567
#>  [3,] 1.07237480 0.65723045 0.2510275106
#>  [4,] 0.82179955 1.97845408 0.0006960564
#>  [5,] 0.16948180 0.20847869 0.4185536221
#>  [6,] 0.06798135 0.52060777 1.5060219435
#>  [7,] 0.07903918 0.07903918 0.0790391764
#>  [8,] 0.11481490 0.11481490 0.1148149012
#>  [9,] 0.18986521 0.09780169 0.1898652087
#> [10,] 0.26438518 0.34581751 0.9400917160
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.090326696 0.09032670 0.09032670
#>  [2,] 0.158356235 0.16585572 1.15113187
#>  [3,] 0.321585033 0.32158503 0.32158503
#>  [4,] 0.897058373 0.06939549 0.27490330
#>  [5,] 0.238776830 0.07721241 0.25900013
#>  [6,] 0.998031365 0.99803137 0.01646330
#>  [7,] 0.403095725 0.40309572 0.02812142
#>  [8,] 0.002253745 0.01105363 0.01105363
#>  [9,] 0.135356308 0.13535631 0.13535631
#> [10,] 0.425957624 0.05707513 0.05707513
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56331243 0.56331243 0.56331243
#>  [2,] 1.23575976 0.84799356 0.19863017
#>  [3,] 0.31400483 0.19338599 0.19338599
#>  [4,] 1.61611169 0.69011365 0.09961936
#>  [5,] 0.05034041 0.07344221 0.07344221
#>  [6,] 0.32225756 0.54411071 0.43388150
#>  [7,] 0.16335948 0.25309248 0.25309248
#>  [8,] 0.18515902 0.18515902 0.18515902
#>  [9,] 0.28948690 0.28948690 0.23808934
#> [10,] 0.67097062 0.17476103 0.51981127

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>             [,1]      [,2]         [,3]
#>  [1,] 0.11062467 0.6154945 0.1106246728
#>  [2,] 0.92201261 0.9213184 0.1843919163
#>  [3,] 0.34328119 0.3432812 0.3432811860
#>  [4,] 0.35105926 0.3510593 0.1758466951
#>  [5,] 0.38970016 0.1844692 0.1844691611
#>  [6,] 0.14026707 0.2011999 0.0009830542
#>  [7,] 0.39144789 0.5112050 0.3602933421
#>  [8,] 0.47118917 0.2210675 0.1407180840
#>  [9,] 0.81457770 0.4230269 0.8982210337
#> [10,] 0.06481979 0.1146208 0.1146207941
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.63675798 0.59635732 0.37495707
#>  [2,] 0.00164835 0.06276805 0.05808482
#>  [3,] 0.19265855 0.13541263 0.13541263
#>  [4,] 0.47898421 0.51475237 0.51475237
#>  [5,] 0.73036812 0.21227550 0.73036812
#>  [6,] 0.05882601 0.05882601 0.31145982
#>  [7,] 0.07122021 0.33928108 0.51553291
#>  [8,] 0.01933205 0.21372134 0.01933205
#>  [9,] 0.84337945 0.84337945 0.90844781
#> [10,] 0.05374556 0.29245356 0.47059183
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>             [,1]       [,2]         [,3]
#>  [1,] 0.08667851 0.08667851 8.667851e-02
#>  [2,] 0.41820161 0.17158324 8.353956e-02
#>  [3,] 0.06544134 0.25234507 2.523451e-01
#>  [4,] 0.20876592 1.34542845 2.087659e-01
#>  [5,] 0.26320673 0.26320673 2.632067e-01
#>  [6,] 0.02170184 0.13029852 2.170184e-02
#>  [7,] 0.42520479 0.82361660 2.606199e-01
#>  [8,] 0.76832300 1.11048825 8.772934e-05
#>  [9,] 0.53456407 0.53456407 4.333775e-01
#> [10,] 0.37755463 0.08949250 3.775546e-01
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.351778764 0.35177876 0.004139503
#>  [2,] 0.141632955 0.19452934 0.105207308
#>  [3,] 0.184633073 0.08596978 0.085969775
#>  [4,] 0.325031674 0.18695525 0.186955247
#>  [5,] 0.396622070 0.46322149 1.104300891
#>  [6,] 0.121697189 0.53579860 0.069862766
#>  [7,] 0.008372871 0.74341996 0.044927671
#>  [8,] 0.640757052 0.52799600 0.013456055
#>  [9,] 0.077353722 0.07735372 0.077353722
#> [10,] 0.451747300 0.63818153 0.506167546