Evaluate Bernstein Function differences

Calculate the values for a Bernstein function and its higher-order, alternating iterated forward differences, possibly scaled by a binomial coefficient, i.e., $$ \binom{n}{k} {(-1)}^{j-1} \Delta^{j}{ \psi(c x) } , \quad x > 0 . $$ The evaluation of Bernstein functions using this formula is usually not numerically stable. Consequently, the various alternative approaches are used dependent on the class of the Bernstein function.

Usage

valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

Arguments

object

An object deriving from the class BernsteinFunction.

x

A nonnegativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

A nonnegative integer with the order of the alternating iterated forward differences taken on the Bernstein function.

n, k

Nonnegative numbers for the binomial factor.

cscale

A positive number for the composite scaling factor.

...

Pass-through parameter.

See also

Examples

bf <- AlphaStableBernsteinFunction(alpha = 0.7)
valueOf(bf, 1:5)
#> [1] 1.000000 1.624505 2.157669 2.639016 3.085169