Virtual superclass for Bernstein functions

A virtual superclass for all implementations of the various classes of Bernstein functions.

Usage

valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

uexIntensities(object, d, cscale = 1, ...)

intensities(object, d, cscale = 1, ...)

exIntensities(object, d, cscale = 1, ...)

exQMatrix(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
uexIntensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
exIntensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
intensities(object, d, cscale = 1, ...)

# S4 method for BernsteinFunction
exQMatrix(object, d, cscale = 1, ...)

Arguments

object

An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

x

a nonnegativ numeric vector at which the iterated difference of the Bernstein function is evaluated.

difference_order

a nonnegative integer with the order of the alternating iterated forward differences taken on the Bernstein function.

n, k

nonnegative numbers for the binomial factor.

cscale

a positive numeric scalar with the composite scaling factor.

...

pass-through parameter

d

a positive integer, larger than two, for the dimension.

Details

A Bernstein function is a nonnegative, nondecreasing, infinitely often differentiable function with whose recursive finite forward differences have alternating signs: $$ {(-1)}^{i-1} \Delta^{i}{ \psi{(x)} } \geq 0 , \quad \forall i \in \mathbb{N}, x \geq 0 . $$

Extendible Marshall--Olkin distributions

A Bernstein function an extendible Marshall–Olkin distribution with exchangeable shock-size arrival intensities $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta^{i}{ \psi{(d-i)} } , \quad 1 \leq i \leq d . $$

The (upper triagonal) infinitesimal Markov generator of the associated death-counting process is calculated recursively: $$ q_{0, i}^\ast = \eta_{i} , \quad i \in {\{ 1 , \ldots , d \}} , $$ and $$ q_{i+1, j+1}^\ast = \frac{d-j}{d-i} q_{i,j}^\ast + \frac{j+1-i}{d-i} q_{i, j+1}^\ast , \quad 0 \leq i < j \leq d . $$

Functions

valueOf(): Calculate the values for a Bernstein function and its higher-order, alternating iterated forward differences, possibly scaled by a binomial coefficient, i.e., $$ {(-1)}^{j-1} \Delta^{j}{ \psi(c x) } , \quad x > 0 . $$

uexIntensities(): Calculates (unscaled) exchangeable shock-arrival intensities, see rmo() and rexmo().

intensities(): Calculates the shock-arrival intensities, the intensities parameter for rmo().

exIntensities(): Calculates exchangeable shock-size-arrival intensities, the ex_intensities parameter for rexmo().

exQMatrix(): Calculates the infinitesimal Markov generator matrix of the corresponding (Markovian) default-counting process, used internally by rexmo().

References

Schilling RL, Song R, Vondracek Z (2012). Bernstein functions, 2 edition. De Gruyter. doi:10.1515/9783110269338 .

See also

Examples

exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.6970130 0.2615107 0.5933218

exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.6970130 0.2615107 0.5933218

intensities(AlphaStableBernsteinFunction(4e-1), 3L)
#> [1] 0.23233766 0.23233766 0.08717025 0.23233766 0.08717025 0.08717025 0.59332184

exQMatrix(AlphaStableBernsteinFunction(4e-1), 3L)
#>           [,1]      [,2]       [,3]      [,4]
#> [1,] -1.551846  0.697013  0.2615107 0.5933218
#> [2,]  0.000000 -1.319508  0.6390158 0.6804921
#> [3,]  0.000000  0.000000 -1.0000000 1.0000000
#> [4,]  0.000000  0.000000  0.0000000 0.0000000