Virtual superclass for Bernstein function objects

Bernstein functions parametrize extendible Marshall--Olkin distributions. They are closed under addition, scalar multiplication, composite scalar multiplication, and consequently also convex recombination:

Pass a Bernstein function object to rextmo() to simulate from the associated extendible MO distribution.

Use SumOfBernsteinFunctions for adding two Bernstein functions,

Use ScaledBernsteinFunction for scalar multiplication of a Bernstein function,

Use CompositeScaledBernsteinFunction for composite scalar multiplication of a Bernstein function, and

Use ConvexCombinationOfBernsteinFunctions for convex recombination of Bernstein functions.

Details

Mathematical definition

A Bernstein function is a nonnegative, nondecreasing, infinitely often differentiable function whose recursive finite forward differences have alternating signs: $$ {(-1)}^{i-1} \Delta^{i}{ \psi{(x)} } \geq 0 , \quad \forall i \in \mathbb{N}, x \geq 0 . $$

Lévy-Khintchine representation

Bernstein functions have the so-called Lévy-Khintchine representation: $$ \psi{(x)} = a x + b + \int_{0}^{\infty}{ {\left[ 1 - e^{-x y} \right]} {\nu{(dy)}} } , \quad x \geq 0 , $$ for some nonnegative constants \(a\) and \(b\) and a Lévy measure \(\nu\). A Lévy measure is a measure on the Borel sets of the nonnegative real line that satisfies the following properties: $$ \int_{0}^{\infty}{ \min{\{ 1 , y \}} {\nu{(dy)}} < \infty . } $$

Lévy-frailty model and ext. MO distributions

Bernstein functions are uniquely linked to extendible Marshall--Olkin distributions via the Lévy-frailty model.

First, each Bernstein functions is uniquely linked to a Lévy subordinator via the Lévy-Khintchine representation: $$ \Lambda{(t)} = \begin{cases} b t + \Lambda_{\nu}{(t)} , & \text{if } t < \epsilon , \\ \infty , & \text{otherwise} , \end{cases} $$ where \(\Lambda_{\nu}{(t)}\) is the Lévy subordinator associated with the Lévy measure \(\nu\) and \(\epsilon\) is an independently exponentially distributed random variable with rate \(a\). If \(\nu\) is a finite measure, the Lévy subordinator is a compound Poisson process with intensity \(\nu{(0, \infty)}\) and jump sizes \(\nu{(dy)} / \nu{((0, \infty))} \).

Second, the Lévy subordinator is unique linked to an extendible Marshall--Olkin distribution via the stochastic representation: $$ \tau_{i} = \inf{\left \{ t \geq 0 : \Lambda{(t)} > E_{i} \right \}} , \quad 1 \leq i \leq d , $$ for independently unit exponentially distributed random variables \(E_{i}\).

Ext. MO parameter interpretation

The Lévy-frailty model motivates the following interpretation of the parameters for ext. MO distributions:

a is the killing rate,

b is the drift,

gamma is a scaling factor for the total shock-arrival intensity,

family is the name of the pure-jump Lévy measure, and

eta are the pure-jump family parameters.

To understand the influence of these parameters on the extendible Marshall–Olkin distribution's dependence properties, the following considerations are helpful:

A pure-killing Bernstein function (i.e., \(a > 0\), \(b = 0\), and \(\nu \equiv 0\)) corresponds to complete comononotonicity.

A pure-drift Bernstein function (i.e., \(a = 0\), \(b > 0\), and \(\nu \equiv 0\)) corresponds to independence.

A pure-jump Bernstein function (i.e., \(a = 0\), \(b = 0\), and \(\nu \not\equiv 0\)) can model various dependence structures. However, larger jump intensities lead to weaker dependendence and larger jump sizes lead to stronger dependence.

Consequently, weighting these cases with the parameters a, b, and gamma allows for a flexible modeling of the dependence structure.

Ext.MO marginal and dependence properties

For a given Bernstein function, the marginal rate and lower-tail dependence coefficient of the associated extendible Marshall–Olkin distribution can be calculated using as follows: $$ \text{Marginal rate} = \psi{(1)} $$ and $$ \text{LTDC} = 2 - \psi{(2)} / \psi{(1)} . $$

Evaluating Bernstein functions

In context of extendible Marshall--Olkin distributions, the following expression is frequently evaluated: $$ \binom{n}{k} {(-1)}^{j-1} \Delta^{j}{ \psi{( c x )} } , \quad 0 \leq k \leq n , j \in \mathbb{N}, x \geq 0 . $$ The evaluation of Bernstein functions using this formula is usually not numerically stable. Consequently, the various alternative approaches are used dependent on the class of the Bernstein function. Use the method valueOf() to evaluate or approximate this expression for a given Bernstein function.

Exchangeable Marshall–Olkin distributions

An alternative stochastic representation of an exchangeable Marshall–Olkin distributions is given by the so-called Markovian death-counting model. It defines the components' death times as randomized order statistics simulated via the Markovian death-counting processes with infinitesimal generator matrix: $$ q_{i, j}^\ast = \binom{d-i}{j-i} \begin{cases} -\psi{(d-i)} , & \text{if } i = j , \\ {(-1)}^{j-i-1} \Delta^{j-i}{ \psi{(d-i)} } , & \text{if } i < j , \\ 0 , & \text{otherwise} . \end{cases} $$ The evaluation of the infinitesimal generator matrix using this formula is usually not numerically stable. Consequently, the various alternative approaches are used dependent on the class of the Bernstein function. Use the method exQMatrix() to evaluate or approximate this expression for a given Bernstein function.

For the all-alive-state, the generator's first row has the interpretation of exchangeable shock-size-arrival intensities: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta^{i}{ \psi{(d-i)} } , \quad 1 \leq i \leq d . $$ As noted above, their evaluation is usually not numerically stable, and various alternative approaches are used dependent on the class of the Bernstein function. Use the method exIntensities() to evaluate or approximate them.

The Exogeneous shock model and the Arnold model

Another alternative stochastic representation of Marshall–Olkin distributions is

References

Schilling RL, Song R, Vondracek Z (2012). Bernstein functions, 2 edition. De Gruyter. doi:10.1515/9783110269338 . Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also