Virtual superclass for complete Bernstein functions

A virtual superclass for all Bernstein functions which can represented by a Stieltjes density (no drift or killing rate). That means that there exists a Stieltjes measure \(\sigma\) such that $$ \psi(x) = \int_0^\infty \frac{x}{x + u} \sigma(du) , x > 0 . $$

Details

Evaluation of Complete Bernstein functions

For continuous Stieltjes densities, the values of the Bernstein function are calculated with stats::integrate() by using the representation $$ \psi(x) = \int_{0}^{\infty} x \mathrm{Beta}(1, x + u) \sigma(du), \quad x > 0 , $$ and the values of the iterated differences are calculated by using the representation $$ (-1)^{j-1} \Delta^{j} \psi(x) = \int_{0}^{\infty} u \mathrm{Beta}(j+1, x + u) \sigma(du) , \quad x > 0 . $$

For discrete Lévy densities \(\sigma(du) = \sum_{i} y_i \delta_{u_i}(du)\), the values of the Bernstein function are calculated by using the representation $$ \psi(x) = \sum_{i} x \mathrm{Beta}(1, x + u_i) y_i, \quad x > 0 , $$ and the values of the iterated differences are calculated by using the representation $$ (-1)^{j-1} \Delta^{j} \psi(x) = \sum_{i} u_i \mathrm{Beta}(j+1, x + u_i) y_i , \quad x > 0 . $$

See also