Class for Pareto Bernstein functions

For the Pareto-jump compound Poisson process with index \(0 < \alpha < 1\) and cutoff point \(x0\), the corresponding Bernstein function is $$ \psi(x) = 1 - e^{-x x_0} + (x_0 x)^\alpha \Gamma(1-\alpha, x_0 x) , x>0 . $$

Usage

# S4 method for ParetoBernsteinFunction
initialize(.Object, alpha, x0)

# S4 method for ParetoBernsteinFunction
show(object)

# S4 method for ParetoBernsteinFunction
levyDensity(object)

Arguments

.Object

An object: see the “Initialize Methods” section.

alpha

Positive number between zero and one (bounds excl.).

x0

Positive number.

object

An object deriving from LevyBernsteinFunction (for levyDensity) or CompleteBernsteinFunction (for stieltjesDensity).

Details

For this Bernstein function, the higher-order alternating iterated forward differences are known in closed form but cannot be evaluated numerically without the danger of loss of significance. But we can use numerical integration (here: stats::integrate()) to approximate it with the following representation: $$ {(-1)}^{k-1} \Delta^k \psi(x) = \int_{x_0}^\infty e^{-ux} (1-e^{-u})^k \alpha \frac{{x_0}^\alpha}{t^{1+\alpha}} du, x>0, k>0 . $$

The Pareto Bernstein function, in combination with a linear Bernstein function can be used to approximate the Bernstein function of an \(\alpha\)-stable Subordinator, see Sec. 5.3 of (Loroño 2015) .

Functions

initialize(ParetoBernsteinFunction): Constructor

show(ParetoBernsteinFunction): Display the object.

levyDensity(ParetoBernsteinFunction): see LevyBernsteinFunction

Slots

alpha

The index \(\alpha\)

x0

The cutoff point \(x_0\)

Lévy density

$$ \nu(du) = \alpha \frac{x_0^\alpha}{u^{\alpha + 1}}, \quad u > x_0 . $$

References

Loroño LF (2015). Selected topics in financial engineering: first-exit times and dependence structures of Marshall-Olkin Kind. Ph.D. thesis, University of the Basque Country.

See also

Examples

ParetoBernsteinFunction()
#> An object of class "ParetoBernsteinFunction"
#> 	 (invalid or not initialized)
ParetoBernsteinFunction(alpha = 0.2, x0 = 1)
#> An object of class "ParetoBernsteinFunction"
#> - alpha: 0.2
#> - x0: 1