Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.67445668 1.17280792 1.67445668
#>  [2,] 0.06427443 0.06427443 0.06427443
#>  [3,] 0.89391017 0.89391017 0.82048075
#>  [4,] 0.86744945 0.34122463 0.86744945
#>  [5,] 0.74600989 0.70546794 0.05443399
#>  [6,] 0.90759038 0.06991403 0.06991403
#>  [7,] 1.07312987 1.58501416 1.97539776
#>  [8,] 0.40875070 0.60755659 0.60755659
#>  [9,] 0.21276319 0.68490900 0.21276319
#> [10,] 0.54931246 0.54931246 0.54931246
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>             [,1]       [,2]      [,3]
#>  [1,] 1.31531160 0.35382771 0.2338608
#>  [2,] 0.39009136 0.08810442 0.1939502
#>  [3,] 2.72297980 0.26725583 1.3172751
#>  [4,] 0.82012128 1.98967423 0.1408802
#>  [5,] 0.12753624 0.05816473 0.1964188
#>  [6,] 1.00898885 0.82383231 0.7258054
#>  [7,] 0.10008590 0.82345975 0.9465018
#>  [8,] 0.01475362 0.23392923 0.3342017
#>  [9,] 0.07226406 0.64670545 0.1373984
#> [10,] 1.34800689 1.07352989 0.6350795
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06718105 0.06718105 0.06718105
#>  [2,] 0.70412488 0.70412488 0.70412488
#>  [3,] 1.20308216 1.20308216 1.20308216
#>  [4,] 0.22017041 0.22017041 0.22017041
#>  [5,] 1.30089347 1.30089347 1.30089347
#>  [6,] 0.09513753 0.09513753 0.09513753
#>  [7,] 0.36939310 0.36939310 0.36939310
#>  [8,] 0.12106060 0.12106060 0.12106060
#>  [9,] 3.59497443 3.59497443 3.59497443
#> [10,] 0.11490137 0.11490137 0.11490137

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.5386140 0.2941105 0.3214207
#>  [2,] 0.7154904 0.7154904 0.3015805
#>  [3,] 0.8697815 0.7097660 0.9409267
#>  [4,] 0.1431254 0.1431254 0.1431254
#>  [5,] 0.3779328 0.3779328 0.3779328
#>  [6,] 1.5192692 1.5020849 0.1827640
#>  [7,] 1.1403037 1.2823590 1.1403037
#>  [8,] 1.1824302 1.1824302 1.1824302
#>  [9,] 0.2043407 0.7304778 0.7304778
#> [10,] 0.5002273 0.5002273 0.4378863
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6675288 0.33405944 2.52626044
#>  [2,] 1.2839479 1.00163436 1.43404135
#>  [3,] 1.1673698 0.30910161 2.02472309
#>  [4,] 0.3266860 1.97927515 0.50798129
#>  [5,] 1.2182733 0.73860133 0.62971403
#>  [6,] 0.1181400 0.58040823 2.08441059
#>  [7,] 2.8771663 1.01583355 1.38885770
#>  [8,] 2.4033471 2.53794340 1.78829662
#>  [9,] 3.1800642 0.07024498 0.08346382
#> [10,] 0.6482357 0.15652210 0.41755230
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41551762 0.41551762 0.41551762
#>  [2,] 0.11040336 0.11040336 0.11040336
#>  [3,] 0.60963486 0.60963486 0.60963486
#>  [4,] 3.08639491 3.08639491 3.08639491
#>  [5,] 0.01381212 0.01381212 0.01381212
#>  [6,] 1.46455780 1.46455780 1.46455780
#>  [7,] 0.64735394 0.64735394 0.64735394
#>  [8,] 0.15470242 0.15470242 0.15470242
#>  [9,] 0.35025917 0.35025917 0.35025917
#> [10,] 0.59165018 0.59165018 0.59165018

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9853425 1.4902779 0.5430459
#>  [2,] 1.5529501 1.5529501 0.5504504
#>  [3,] 1.9067629 1.9067629 0.3869912
#>  [4,] 0.7234934 0.7234934 0.7234934
#>  [5,] 0.1197150 2.7593503 0.6605526
#>  [6,] 0.1373801 1.2489560 0.1373801
#>  [7,] 0.2993108 0.2993108 0.2993108
#>  [8,] 0.4318750 0.4318750 0.4318750
#>  [9,] 2.7837024 1.4674408 2.8976524
#> [10,] 0.6435889 0.6435889 0.6435889
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43963345 0.57019972 0.79262445
#>  [2,] 2.27279045 0.05578051 0.59059116
#>  [3,] 0.41607809 0.06190309 0.74796532
#>  [4,] 0.34900627 2.49457480 2.52631974
#>  [5,] 0.30873165 0.87756197 1.49165356
#>  [6,] 2.42028142 1.60051424 0.02863809
#>  [7,] 0.98010450 1.85866807 1.47479371
#>  [8,] 0.92466462 2.60017068 2.18634159
#>  [9,] 0.02178152 2.94299654 1.19546915
#> [10,] 0.09734178 0.16625687 0.36741469
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58435306 0.58435306 0.58435306
#>  [2,] 2.42332265 2.42332265 2.42332265
#>  [3,] 0.13288317 0.13288317 0.13288317
#>  [4,] 1.35580168 1.35580168 1.35580168
#>  [5,] 0.02838263 0.02838263 0.02838263
#>  [6,] 0.79044047 0.79044047 0.79044047
#>  [7,] 0.87626291 0.87626291 0.87626291
#>  [8,] 0.18618239 0.18618239 0.18618239
#>  [9,] 2.18486283 2.18486283 2.18486283
#> [10,] 0.32088959 0.32088959 0.32088959

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5317531 1.5558612 2.2520689
#>  [2,] 0.4107317 0.5242374 0.5242374
#>  [3,] 1.5506232 1.5506232 1.5506232
#>  [4,] 0.3743279 0.3743279 0.3743279
#>  [5,] 1.8371289 0.5655280 2.6524657
#>  [6,] 3.4150544 2.8743633 2.0960533
#>  [7,] 1.6605534 0.5417275 1.6605534
#>  [8,] 1.3617376 0.1389395 0.6973011
#>  [9,] 1.7957040 1.7957040 1.7957040
#> [10,] 1.7927952 0.9573316 1.7727445
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.30104428 1.082099773 0.59985242
#>  [2,] 3.54203440 2.862914332 0.27604260
#>  [3,] 0.01628188 0.329999347 1.78215961
#>  [4,] 1.26194898 1.066109746 2.47221350
#>  [5,] 1.89120790 0.848119765 0.09758546
#>  [6,] 0.59572101 0.774817002 1.58048366
#>  [7,] 0.31451600 0.005158928 1.91521628
#>  [8,] 0.08524468 1.908577103 0.24729293
#>  [9,] 2.12453956 0.401672946 0.40894252
#> [10,] 2.77439668 0.235429025 1.70858909
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1277850 1.1277850 1.1277850
#>  [2,] 0.5936300 0.5936300 0.5936300
#>  [3,] 0.8754027 0.8754027 0.8754027
#>  [4,] 0.3090675 0.3090675 0.3090675
#>  [5,] 1.8597004 1.8597004 1.8597004
#>  [6,] 0.3931107 0.3931107 0.3931107
#>  [7,] 0.7147029 0.7147029 0.7147029
#>  [8,] 0.0866368 0.0866368 0.0866368
#>  [9,] 1.5965072 1.5965072 1.5965072
#> [10,] 0.7177487 0.7177487 0.7177487