Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3785814 1.1237894 0.2704799
#>  [2,] 2.7446094 0.7867448 3.5513331
#>  [3,] 0.4412136 0.6230040 0.4412136
#>  [4,] 1.2493247 1.2464559 1.2464559
#>  [5,] 0.1202796 0.7533109 0.1018503
#>  [6,] 1.2932413 1.3866548 0.6967410
#>  [7,] 0.2563541 1.8691709 2.8329147
#>  [8,] 4.8673999 0.4430588 0.5810045
#>  [9,] 0.3367547 0.3367547 0.4461855
#> [10,] 1.7099990 0.2319876 1.7571374
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>             [,1]      [,2]      [,3]
#>  [1,] 0.06773834 0.2028867 1.3451560
#>  [2,] 0.61028397 0.3129418 0.1488302
#>  [3,] 1.91318269 1.4611179 1.7487597
#>  [4,] 1.35241224 0.3891132 0.2324515
#>  [5,] 0.18578698 0.6254791 0.2631957
#>  [6,] 0.45966760 0.4788949 1.1667235
#>  [7,] 0.13371667 0.4626237 0.6368613
#>  [8,] 0.74927876 0.5753582 1.3575470
#>  [9,] 0.38222817 2.3072891 0.3342881
#> [10,] 0.06236342 1.1541198 0.1142672
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4641659 0.4641659 0.4641659
#>  [2,] 2.5135185 2.5135185 2.5135185
#>  [3,] 0.1088544 0.1088544 0.1088544
#>  [4,] 0.9586277 0.9586277 0.9586277
#>  [5,] 0.7299953 0.7299953 0.7299953
#>  [6,] 0.4719184 0.4719184 0.4719184
#>  [7,] 1.1140933 1.1140933 1.1140933
#>  [8,] 0.4237286 0.4237286 0.4237286
#>  [9,] 0.2664121 0.2664121 0.2664121
#> [10,] 1.1613227 1.1613227 1.1613227

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0348276 2.0348276 0.8820617
#>  [2,] 0.2270734 0.1977298 0.7182060
#>  [3,] 1.4400612 1.6876773 1.4400612
#>  [4,] 2.3072893 1.5615244 0.2217981
#>  [5,] 1.2703950 0.3576167 1.0217578
#>  [6,] 2.0939235 0.5736193 1.7253504
#>  [7,] 1.0089600 2.4089111 0.5588430
#>  [8,] 1.0662752 1.6980030 2.8201491
#>  [9,] 1.5072556 1.3134002 0.5704694
#> [10,] 1.6757642 0.2970420 0.2970420
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>             [,1]      [,2]        [,3]
#>  [1,] 2.72955583 0.2306218 0.750155864
#>  [2,] 1.82246745 0.5898028 1.267010539
#>  [3,] 0.66524508 2.0149135 0.370185235
#>  [4,] 0.03785607 0.4185332 0.292435150
#>  [5,] 0.11720595 0.3558614 0.198769054
#>  [6,] 1.24460586 0.4447038 0.081741209
#>  [7,] 0.32749762 0.7950308 0.006142849
#>  [8,] 0.97003067 2.1080779 0.109980353
#>  [9,] 1.35382866 2.2331843 0.424389211
#> [10,] 1.42018009 0.2677006 1.012707653
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02502946 0.02502946 0.02502946
#>  [2,] 1.88609327 1.88609327 1.88609327
#>  [3,] 3.17116940 3.17116940 3.17116940
#>  [4,] 0.25283860 0.25283860 0.25283860
#>  [5,] 0.22635851 0.22635851 0.22635851
#>  [6,] 2.04977186 2.04977186 2.04977186
#>  [7,] 1.01665514 1.01665514 1.01665514
#>  [8,] 0.09191952 0.09191952 0.09191952
#>  [9,] 0.47713824 0.47713824 0.47713824
#> [10,] 0.02303828 0.02303828 0.02303828

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.54766331 6.99182789 0.04185243
#>  [2,] 0.03328951 0.03328951 0.25885467
#>  [3,] 0.10862132 0.67463915 0.17330023
#>  [4,] 3.23130034 2.37459047 0.21109034
#>  [5,] 0.81902032 0.32571059 0.02634402
#>  [6,] 2.40347651 0.33779348 2.54787110
#>  [7,] 1.29487294 5.42618858 2.23700642
#>  [8,] 0.95797798 0.24989022 0.04083838
#>  [9,] 0.23849619 0.23849619 0.23849619
#> [10,] 2.75306307 0.04150142 2.76535177
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>            [,1]       [,2]        [,3]
#>  [1,] 1.3024797 0.60580863 0.624797987
#>  [2,] 0.2250360 0.07737299 0.464428571
#>  [3,] 1.7503766 3.23867516 0.289489192
#>  [4,] 0.1973391 0.40378989 1.201628121
#>  [5,] 0.5667094 0.66023599 2.072635321
#>  [6,] 0.6932684 2.64030969 0.001112219
#>  [7,] 0.6699277 0.11900902 0.099599971
#>  [8,] 0.6519861 0.31945303 0.031128903
#>  [9,] 1.6496666 1.91064644 0.239379556
#> [10,] 4.2644909 1.46918982 0.504625843
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.21445411 1.21445411 1.21445411
#>  [2,] 1.47284663 1.47284663 1.47284663
#>  [3,] 0.37711122 0.37711122 0.37711122
#>  [4,] 1.12298387 1.12298387 1.12298387
#>  [5,] 0.54109796 0.54109796 0.54109796
#>  [6,] 0.90851410 0.90851410 0.90851410
#>  [7,] 0.29870535 0.29870535 0.29870535
#>  [8,] 0.08818292 0.08818292 0.08818292
#>  [9,] 2.53536766 2.53536766 2.53536766
#> [10,] 0.09709676 0.09709676 0.09709676

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8438665 0.5772495 0.8438665
#>  [2,] 0.3030586 0.3030586 0.3030586
#>  [3,] 0.4560608 0.4560608 0.4990515
#>  [4,] 0.5414098 0.9511718 0.2130009
#>  [5,] 0.3737830 0.3737830 0.3737830
#>  [6,] 0.3024302 0.6567318 2.0192695
#>  [7,] 0.3167496 0.6706243 0.6706243
#>  [8,] 0.7306647 0.9757889 0.1109966
#>  [9,] 0.0151173 0.0151173 0.0151173
#> [10,] 1.5660220 0.4370221 1.5660220
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.1557123 0.2547367 0.58123301
#>  [2,] 3.8363664 1.1610269 0.00606526
#>  [3,] 0.6762820 0.5663163 2.09852477
#>  [4,] 0.2565571 1.7593424 0.19149026
#>  [5,] 3.5432772 0.4354382 0.99891978
#>  [6,] 1.5383334 0.7041179 0.06122514
#>  [7,] 0.7480064 0.4937856 0.23955998
#>  [8,] 0.9799922 0.6294108 0.26558582
#>  [9,] 0.1028940 0.1759678 0.07421237
#> [10,] 0.3827953 2.0411761 0.64868321
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.0340036 0.0340036 0.0340036
#>  [2,] 0.1998899 0.1998899 0.1998899
#>  [3,] 1.6272280 1.6272280 1.6272280
#>  [4,] 0.7460372 0.7460372 0.7460372
#>  [5,] 0.7746036 0.7746036 0.7746036
#>  [6,] 1.4555228 1.4555228 1.4555228
#>  [7,] 1.1636233 1.1636233 1.1636233
#>  [8,] 0.5448310 0.5448310 0.5448310
#>  [9,] 0.1077834 0.1077834 0.1077834
#> [10,] 0.5051110 0.5051110 0.5051110