Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 2.39111644 0.68185810 0.31705714
#>  [2,] 0.28038976 1.92146816 0.28038976
#>  [3,] 2.69259296 1.55958484 2.69259296
#>  [4,] 0.07197897 0.07197897 0.07197897
#>  [5,] 0.40091526 0.40091526 0.40091526
#>  [6,] 1.52752762 0.78532714 1.52752762
#>  [7,] 0.79229382 0.79229382 0.60798351
#>  [8,] 1.69301004 1.72751671 1.72751671
#>  [9,] 0.08113041 0.08113041 0.08113041
#> [10,] 0.31578303 0.31578303 0.31296502
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8029063 0.5409055 0.8764946
#>  [2,] 0.7924643 4.5682042 0.2012464
#>  [3,] 1.8289174 1.9516222 1.6156468
#>  [4,] 0.1069293 1.0873210 0.5569240
#>  [5,] 1.0988145 2.4446356 0.7046535
#>  [6,] 0.2436393 0.4874676 0.8331682
#>  [7,] 2.8446338 2.2666462 0.3895147
#>  [8,] 1.3449809 2.4007057 0.8400421
#>  [9,] 1.0822319 0.2197074 2.3249398
#> [10,] 0.4368697 0.4708971 0.2119699
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7555515 0.7555515 0.7555515
#>  [2,] 0.4086169 0.4086169 0.4086169
#>  [3,] 1.4142877 1.4142877 1.4142877
#>  [4,] 1.4887606 1.4887606 1.4887606
#>  [5,] 2.4338885 2.4338885 2.4338885
#>  [6,] 1.9969365 1.9969365 1.9969365
#>  [7,] 0.9501088 0.9501088 0.9501088
#>  [8,] 3.8255672 3.8255672 3.8255672
#>  [9,] 0.3733719 0.3733719 0.3733719
#> [10,] 1.0407706 1.0407706 1.0407706

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4462931 0.4462931 0.4462931
#>  [2,] 0.7658093 0.7739609 2.0663409
#>  [3,] 3.6743670 1.5120243 2.7679600
#>  [4,] 1.2235394 1.5051443 3.6544112
#>  [5,] 3.4452348 1.6019861 0.5434982
#>  [6,] 0.7109350 0.7109350 0.7109350
#>  [7,] 0.4393410 0.4393410 2.0869523
#>  [8,] 0.6778149 0.1700424 1.6696338
#>  [9,] 2.4763616 2.4763616 0.0624061
#> [10,] 1.0543721 1.0543721 0.7520343
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 1.9202298 0.84825068 1.9248076
#>  [2,] 1.8346451 0.36983880 0.7015965
#>  [3,] 1.1992566 0.07237578 3.1975111
#>  [4,] 0.8731173 1.06272855 0.2449838
#>  [5,] 1.0601465 0.35120923 0.4613831
#>  [6,] 0.1929532 0.96339962 3.0707279
#>  [7,] 1.8674713 0.04592372 1.4166183
#>  [8,] 0.4205749 2.86584526 0.1240178
#>  [9,] 1.1923864 0.60587283 2.7635893
#> [10,] 0.3013562 0.73974029 0.1718825
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6952517 1.6952517 1.6952517
#>  [2,] 0.6125562 0.6125562 0.6125562
#>  [3,] 0.2595634 0.2595634 0.2595634
#>  [4,] 2.2925100 2.2925100 2.2925100
#>  [5,] 1.0017830 1.0017830 1.0017830
#>  [6,] 0.2959366 0.2959366 0.2959366
#>  [7,] 0.4373159 0.4373159 0.4373159
#>  [8,] 0.8232866 0.8232866 0.8232866
#>  [9,] 0.2313076 0.2313076 0.2313076
#> [10,] 1.8044942 1.8044942 1.8044942

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7061794 4.4952000 2.2704935
#>  [2,] 1.3648103 1.3648103 1.3648103
#>  [3,] 1.0337540 1.0337540 1.4627693
#>  [4,] 0.1851425 1.2208655 0.5079268
#>  [5,] 0.1282614 0.7358506 1.7578921
#>  [6,] 0.7356369 0.7356369 0.7356369
#>  [7,] 0.1882838 0.1882838 0.9248593
#>  [8,] 0.1685830 0.4269816 0.1685830
#>  [9,] 3.6320211 0.7135990 2.1516340
#> [10,] 0.2487274 0.7844573 0.8584547
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>              [,1]       [,2]      [,3]
#>  [1,] 0.004495364 0.13008661 1.1985473
#>  [2,] 1.036747054 0.15975721 2.6096924
#>  [3,] 0.234416201 0.05356609 0.4299915
#>  [4,] 2.818087458 2.94754981 0.2808761
#>  [5,] 0.893268854 3.00087688 0.3786713
#>  [6,] 3.141142234 0.19085576 0.6141667
#>  [7,] 0.069772250 1.24426072 1.0694838
#>  [8,] 0.794867898 0.35678822 0.8004856
#>  [9,] 0.370726461 0.62422827 0.5947007
#> [10,] 0.105091332 1.05072226 0.5824460
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84327084 0.84327084 0.84327084
#>  [2,] 0.11427070 0.11427070 0.11427070
#>  [3,] 1.99675611 1.99675611 1.99675611
#>  [4,] 0.05188269 0.05188269 0.05188269
#>  [5,] 0.64993582 0.64993582 0.64993582
#>  [6,] 0.25902187 0.25902187 0.25902187
#>  [7,] 1.05754155 1.05754155 1.05754155
#>  [8,] 0.48786105 0.48786105 0.48786105
#>  [9,] 0.27850959 0.27850959 0.27850959
#> [10,] 1.28187813 1.28187813 1.28187813

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>              [,1]       [,2]      [,3]
#>  [1,] 0.689792163 0.02368449 1.5323948
#>  [2,] 0.998734283 0.01981413 0.4540969
#>  [3,] 0.044994475 0.04865080 0.0486508
#>  [4,] 0.390625012 0.28589531 0.9562663
#>  [5,] 0.007509619 5.42308338 0.4265153
#>  [6,] 1.624750531 1.23413367 1.2341337
#>  [7,] 0.073049687 0.07723713 0.1552334
#>  [8,] 1.359304107 2.11147609 2.1114761
#>  [9,] 1.685562095 3.55422555 0.3407083
#> [10,] 0.182253566 2.71800225 2.6984070
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47036875 0.06212402 0.13989235
#>  [2,] 3.13039102 3.15587104 2.99313499
#>  [3,] 0.52198626 1.26942433 0.02997756
#>  [4,] 0.94923971 0.89756259 0.34953456
#>  [5,] 0.08376477 0.29156498 0.45553788
#>  [6,] 0.50521716 0.28534612 0.02364751
#>  [7,] 0.09870081 0.73676987 0.07185337
#>  [8,] 0.92571746 0.55020281 0.41424695
#>  [9,] 0.28962617 0.05979016 0.82073492
#> [10,] 2.75234349 0.58368562 0.52899988
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11975066 0.11975066 0.11975066
#>  [2,] 0.06072977 0.06072977 0.06072977
#>  [3,] 0.10006347 0.10006347 0.10006347
#>  [4,] 0.30080419 0.30080419 0.30080419
#>  [5,] 1.76648932 1.76648932 1.76648932
#>  [6,] 0.04738627 0.04738627 0.04738627
#>  [7,] 1.42444407 1.42444407 1.42444407
#>  [8,] 0.73203713 0.73203713 0.73203713
#>  [9,] 0.44572993 0.44572993 0.44572993
#> [10,] 0.37538676 0.37538676 0.37538676