Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationship of exchangeable shock-size arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-counting model

The Markovian death-counting model simulates the random vector's death-counting process, which is a Markov process, until all components are dead. This defines an order statistic that is used to obtain a sample by a random permutation; see (Sloot 2022) and (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7592804 0.46955595 3.2957286
#>  [2,] 0.1099769 0.10997690 0.1099769
#>  [3,] 0.2324139 1.95008839 0.2324139
#>  [4,] 0.7930417 0.28241620 0.2824162
#>  [5,] 0.5280804 0.52808043 0.5280804
#>  [6,] 0.1715973 0.68039347 0.4518182
#>  [7,] 2.7578379 1.64585866 0.8812193
#>  [8,] 1.4439829 0.62257822 0.2638873
#>  [9,] 0.9785194 1.11762926 2.2461827
#> [10,] 1.0557107 0.06623165 2.1671974
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4356209 1.3051332 0.37749421
#>  [2,] 0.5398070 0.2363022 0.09418562
#>  [3,] 0.3660650 0.4114661 0.48862689
#>  [4,] 1.1031372 0.3522785 0.43495378
#>  [5,] 2.9461240 0.1932220 0.21493515
#>  [6,] 0.7698752 0.6366983 0.38902195
#>  [7,] 0.9634438 2.4707796 2.26836084
#>  [8,] 1.7233935 0.3564566 0.74500661
#>  [9,] 0.1365247 0.4014418 0.03378446
#> [10,] 5.2950858 0.7284916 4.69269187
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.45426091 1.45426091 1.45426091
#>  [2,] 0.18770086 0.18770086 0.18770086
#>  [3,] 0.04316363 0.04316363 0.04316363
#>  [4,] 0.37472482 0.37472482 0.37472482
#>  [5,] 1.32954101 1.32954101 1.32954101
#>  [6,] 4.42650365 4.42650365 4.42650365
#>  [7,] 0.31213191 0.31213191 0.31213191
#>  [8,] 0.43733958 0.43733958 0.43733958
#>  [9,] 0.08954413 0.08954413 0.08954413
#> [10,] 0.78690865 0.78690865 0.78690865

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04912539 0.72777843 0.72777843
#>  [2,] 0.03523465 0.03523465 0.52773440
#>  [3,] 0.53448148 0.18474955 1.35144674
#>  [4,] 0.95764754 1.33796519 0.17502648
#>  [5,] 2.11049838 2.11049838 5.65253155
#>  [6,] 0.15462170 0.15462170 1.92461725
#>  [7,] 0.07535969 1.11817274 0.07535969
#>  [8,] 1.56219171 1.01314727 1.57212851
#>  [9,] 0.45816838 1.54371229 1.33903418
#> [10,] 0.81295642 0.81295642 0.39283386
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.76781092 1.89336828 0.02896769
#>  [2,] 0.98300226 0.39087855 0.67059669
#>  [3,] 0.44105853 2.49262335 1.32590346
#>  [4,] 0.23796976 1.06046255 0.18458448
#>  [5,] 0.01063648 0.98071907 0.38347814
#>  [6,] 0.23573843 0.01766145 0.90594960
#>  [7,] 0.70675279 2.75245135 0.42535809
#>  [8,] 1.55901211 2.30592383 0.53899241
#>  [9,] 0.31859213 0.09079169 0.01791430
#> [10,] 0.45076627 0.17233449 1.70022763
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1683838 1.1683838 1.1683838
#>  [2,] 1.5254088 1.5254088 1.5254088
#>  [3,] 0.8640073 0.8640073 0.8640073
#>  [4,] 0.8316041 0.8316041 0.8316041
#>  [5,] 0.7035932 0.7035932 0.7035932
#>  [6,] 0.5299297 0.5299297 0.5299297
#>  [7,] 1.7275861 1.7275861 1.7275861
#>  [8,] 0.9239711 0.9239711 0.9239711
#>  [9,] 0.1878293 0.1878293 0.1878293
#> [10,] 0.9925791 0.9925791 0.9925791

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07634567 1.07634567 0.95401408
#>  [2,] 0.62388140 0.62388140 0.62388140
#>  [3,] 0.42706201 0.93733692 0.42706201
#>  [4,] 0.23216355 0.23216355 0.23216355
#>  [5,] 0.25276763 0.46745041 0.84491264
#>  [6,] 0.75513832 0.40301399 0.75513832
#>  [7,] 0.06253784 0.06253784 0.06253784
#>  [8,] 1.19076855 0.68587075 0.44522596
#>  [9,] 3.32009679 1.07060864 1.07060864
#> [10,] 2.60744065 0.63476231 0.74975213
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 3.3495560 0.19534857 0.6213569
#>  [2,] 0.9020097 0.07787802 0.2000250
#>  [3,] 0.4292579 0.10457257 2.5892822
#>  [4,] 0.2811996 0.06908237 0.1941377
#>  [5,] 0.3522387 0.41728759 0.4787032
#>  [6,] 0.6785561 1.29266647 0.7428717
#>  [7,] 1.8241437 0.64742383 3.4805860
#>  [8,] 0.2915405 1.11047682 1.5614163
#>  [9,] 0.1142091 1.07017910 5.5361726
#> [10,] 1.2026214 0.11773729 0.2118641
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.63900476 0.63900476 0.63900476
#>  [2,] 2.42017812 2.42017812 2.42017812
#>  [3,] 0.39913162 0.39913162 0.39913162
#>  [4,] 0.25782293 0.25782293 0.25782293
#>  [5,] 2.68073049 2.68073049 2.68073049
#>  [6,] 0.86426390 0.86426390 0.86426390
#>  [7,] 1.30642800 1.30642800 1.30642800
#>  [8,] 0.69639261 0.69639261 0.69639261
#>  [9,] 0.05170849 0.05170849 0.05170849
#> [10,] 0.18912463 0.18912463 0.18912463

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.42654223 2.0115272 2.01152716
#>  [2,] 0.57459975 2.1351687 2.13516870
#>  [3,] 1.83521030 1.6778676 1.67786759
#>  [4,] 1.55005710 2.6161767 1.79301658
#>  [5,] 0.16312257 0.1631226 0.16312257
#>  [6,] 0.44037744 0.4403774 0.44037744
#>  [7,] 1.30768523 1.3076852 1.30768523
#>  [8,] 0.78623100 0.7862310 0.78623100
#>  [9,] 0.07103116 2.4759815 0.07103116
#> [10,] 0.44473877 0.0863428 0.22824434
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1674734 2.95876089 3.2871420
#>  [2,] 0.6899637 1.48778691 2.1598328
#>  [3,] 1.1608679 0.01934236 0.5453775
#>  [4,] 0.2540783 0.77511997 1.1079036
#>  [5,] 0.3998643 0.90802385 2.1559495
#>  [6,] 1.5690225 1.03600576 1.5539752
#>  [7,] 1.3264644 2.01053101 1.3060081
#>  [8,] 1.5797326 0.33234663 0.1186057
#>  [9,] 2.3341878 0.61267054 0.2647717
#> [10,] 2.1351985 0.31916168 0.1140432
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35841748 0.35841748 0.35841748
#>  [2,] 0.08147809 0.08147809 0.08147809
#>  [3,] 0.71618173 0.71618173 0.71618173
#>  [4,] 0.26972791 0.26972791 0.26972791
#>  [5,] 2.52848431 2.52848431 2.52848431
#>  [6,] 2.04416509 2.04416509 2.04416509
#>  [7,] 0.61451090 0.61451090 0.61451090
#>  [8,] 1.26057672 1.26057672 1.26057672
#>  [9,] 1.65760829 1.65760829 1.65760829
#> [10,] 0.60846388 0.60846388 0.60846388