Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 3.67340436 3.67340436 2.77318198
#>  [2,] 1.02308408 1.02308408 1.02308408
#>  [3,] 0.73663467 0.73663467 0.73663467
#>  [4,] 1.61902830 0.25081453 1.84897871
#>  [5,] 0.04306444 0.67739761 0.04306444
#>  [6,] 2.59660582 0.80884319 0.87177806
#>  [7,] 1.97168912 1.97168912 1.51010670
#>  [8,] 0.56203300 0.09401252 0.17295726
#>  [9,] 0.94365865 0.94365865 0.94365865
#> [10,] 0.61983296 0.95233446 1.39124384
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>              [,1]       [,2]      [,3]
#>  [1,] 0.240971246 4.46937196 0.1235226
#>  [2,] 0.006787173 1.05629195 0.2229181
#>  [3,] 5.671608672 0.14427873 1.2212345
#>  [4,] 0.239757222 0.34670052 1.8599170
#>  [5,] 0.287421359 0.88312477 1.0467379
#>  [6,] 2.811082216 0.61291311 0.6674922
#>  [7,] 2.033156713 0.65164645 0.6270563
#>  [8,] 0.574356700 0.06615963 3.0017185
#>  [9,] 2.045374124 1.46201039 2.5531570
#> [10,] 2.261659599 3.91178281 0.2027654
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.89928947 0.89928947 0.89928947
#>  [2,] 1.76092983 1.76092983 1.76092983
#>  [3,] 2.23047729 2.23047729 2.23047729
#>  [4,] 0.69698275 0.69698275 0.69698275
#>  [5,] 0.09771343 0.09771343 0.09771343
#>  [6,] 0.16255200 0.16255200 0.16255200
#>  [7,] 1.13857904 1.13857904 1.13857904
#>  [8,] 3.57672163 3.57672163 3.57672163
#>  [9,] 1.56085482 1.56085482 1.56085482
#> [10,] 0.89949991 0.89949991 0.89949991

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18408595 0.64112461 0.63515373
#>  [2,] 0.36879374 0.07796847 0.07796847
#>  [3,] 0.09486595 0.67812216 2.98518512
#>  [4,] 1.33469479 0.46735809 0.75133076
#>  [5,] 0.66105601 1.20554793 1.20554793
#>  [6,] 0.18726837 0.18726837 0.18726837
#>  [7,] 0.31452009 0.31452009 0.31452009
#>  [8,] 2.07664677 1.71825334 0.46995191
#>  [9,] 0.37646707 2.41737774 1.45809945
#> [10,] 0.40161387 0.76862377 1.28645301
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.2843773 0.12321309 1.81431519
#>  [2,] 0.3525233 0.08504418 0.54959992
#>  [3,] 0.6281024 0.68263762 0.03709487
#>  [4,] 0.3628534 0.83815022 2.96196939
#>  [5,] 0.2039492 0.89474933 0.76106928
#>  [6,] 0.5391263 1.68941776 0.82013636
#>  [7,] 1.9871375 0.45895669 0.33563267
#>  [8,] 0.6757121 0.63375774 0.64868373
#>  [9,] 0.9483049 0.24849546 2.82806291
#> [10,] 1.0963064 0.55312054 1.47438091
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.6518974 2.6518974 2.6518974
#>  [2,] 1.0156948 1.0156948 1.0156948
#>  [3,] 0.5493246 0.5493246 0.5493246
#>  [4,] 1.7059350 1.7059350 1.7059350
#>  [5,] 0.6302063 0.6302063 0.6302063
#>  [6,] 3.5750166 3.5750166 3.5750166
#>  [7,] 1.5522313 1.5522313 1.5522313
#>  [8,] 0.1544477 0.1544477 0.1544477
#>  [9,] 0.7963817 0.7963817 0.7963817
#> [10,] 0.4618057 0.4618057 0.4618057

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.34032732 0.3403273 0.34032732
#>  [2,] 0.42306686 2.0131466 0.92978803
#>  [3,] 0.28073993 0.2807399 0.28073993
#>  [4,] 0.38071555 0.3807155 0.67783361
#>  [5,] 1.37770989 0.4415924 6.01521500
#>  [6,] 1.62301760 1.5817550 1.69596131
#>  [7,] 0.53928619 0.4419029 0.02263727
#>  [8,] 0.08011245 0.1778651 0.17786513
#>  [9,] 1.72880703 0.4898859 0.50787205
#> [10,] 2.09437914 2.0943791 2.35498175
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>            [,1]      [,2]         [,3]
#>  [1,] 0.3362075 0.3430871 3.790794e-06
#>  [2,] 0.7191872 0.0484205 2.344561e-01
#>  [3,] 2.4686841 0.4941912 3.733741e+00
#>  [4,] 0.3089077 1.0216382 8.329017e-01
#>  [5,] 1.3642868 0.2895629 9.444935e-01
#>  [6,] 0.3191339 1.9932380 1.097199e+00
#>  [7,] 0.1098066 1.2871219 3.393740e-01
#>  [8,] 1.7933870 0.9442255 1.584048e+00
#>  [9,] 2.1055384 0.5407486 7.550021e-02
#> [10,] 2.8134022 0.1737126 7.133501e-01
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0939414 1.0939414 1.0939414
#>  [2,] 0.1584764 0.1584764 0.1584764
#>  [3,] 1.4866132 1.4866132 1.4866132
#>  [4,] 0.7348747 0.7348747 0.7348747
#>  [5,] 1.9592114 1.9592114 1.9592114
#>  [6,] 0.8536059 0.8536059 0.8536059
#>  [7,] 0.4840809 0.4840809 0.4840809
#>  [8,] 1.4556650 1.4556650 1.4556650
#>  [9,] 0.3534494 0.3534494 0.3534494
#> [10,] 0.8354802 0.8354802 0.8354802

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.3223221 1.32232210 1.3223221
#>  [2,] 2.1479401 0.07434033 0.4262020
#>  [3,] 1.7231297 1.78873756 1.7887376
#>  [4,] 0.3542314 4.32284602 0.3207955
#>  [5,] 1.1916940 1.19169398 0.7404732
#>  [6,] 0.2139681 0.21396807 2.3206216
#>  [7,] 4.1401718 0.45240930 3.5441748
#>  [8,] 1.5562679 1.55626792 1.5562679
#>  [9,] 0.5426841 1.60792992 0.5426841
#> [10,] 2.3376440 0.27894924 0.5258400
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>               [,1]       [,2]      [,3]
#>  [1,] 0.3144456748 0.16366730 0.3580781
#>  [2,] 0.7928567808 0.82880132 0.0832202
#>  [3,] 0.2762382980 0.01265555 1.8209551
#>  [4,] 2.5110587942 1.09796334 1.0774108
#>  [5,] 0.0002979342 0.55682964 0.8168722
#>  [6,] 0.2366338968 0.12430319 0.9492940
#>  [7,] 1.7345493379 2.78346363 0.9026490
#>  [8,] 0.8153256243 0.30957633 0.9920186
#>  [9,] 0.0569269652 3.63147203 1.0482223
#> [10,] 1.8357017151 0.72926621 0.7437779
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8846848 0.8846848 0.8846848
#>  [2,] 1.3976869 1.3976869 1.3976869
#>  [3,] 0.1266287 0.1266287 0.1266287
#>  [4,] 0.4480026 0.4480026 0.4480026
#>  [5,] 0.9687088 0.9687088 0.9687088
#>  [6,] 1.7708140 1.7708140 1.7708140
#>  [7,] 0.3539017 0.3539017 0.3539017
#>  [8,] 0.2060850 0.2060850 0.2060850
#>  [9,] 1.4512773 1.4512773 1.4512773
#> [10,] 0.2668736 0.2668736 0.2668736