Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>              [,1]        [,2]        [,3]
#>  [1,] 1.306747881 2.311743364 0.862837812
#>  [2,] 1.587506035 2.548691331 1.967823124
#>  [3,] 0.352348154 1.216009809 5.096792911
#>  [4,] 0.179243217 0.179243217 0.179243217
#>  [5,] 0.335714125 0.464501955 0.464501955
#>  [6,] 0.007533603 0.007533603 0.007533603
#>  [7,] 0.474926492 0.137281832 0.474926492
#>  [8,] 0.822488379 2.202162645 1.411510273
#>  [9,] 0.575371505 0.575371505 1.588050027
#> [10,] 1.436181236 1.717769693 1.717769693
rexmo(10, 3, c(3, 0, 0))          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.78655565 0.08910288 1.2812437
#>  [2,] 0.03839550 0.68019796 0.6432592
#>  [3,] 0.61199017 0.18926344 1.1500741
#>  [4,] 3.26642501 0.28991336 0.8242512
#>  [5,] 0.33217419 0.36310221 1.9674307
#>  [6,] 0.07390626 0.89290537 0.2568733
#>  [7,] 0.04666200 0.62217116 2.5821092
#>  [8,] 1.41056595 0.58219826 0.3606327
#>  [9,] 0.29296081 0.22881796 1.8748333
#> [10,] 0.10218720 1.85690355 0.3010903
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5624164 1.5624164 1.5624164
#>  [2,] 0.1286449 0.1286449 0.1286449
#>  [3,] 0.3271123 0.3271123 0.3271123
#>  [4,] 0.1134372 0.1134372 0.1134372
#>  [5,] 0.4517612 0.4517612 0.4517612
#>  [6,] 0.3451201 0.3451201 0.3451201
#>  [7,] 0.8047525 0.8047525 0.8047525
#>  [8,] 0.1791313 0.1791313 0.1791313
#>  [9,] 0.0850397 0.0850397 0.0850397
#> [10,] 1.2199919 1.2199919 1.2199919

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.37929382 0.3792938 0.37929382
#>  [2,] 0.20156839 1.0494875 0.11488978
#>  [3,] 1.78496538 3.5081565 4.02106381
#>  [4,] 0.44632128 0.4463213 0.44632128
#>  [5,] 0.70451874 0.7045187 0.70451874
#>  [6,] 1.83506335 0.8931525 0.71164726
#>  [7,] 0.91161356 0.7300953 0.60561888
#>  [8,] 0.16793409 0.1679341 0.16793409
#>  [9,] 0.69963535 0.6996354 1.50228641
#> [10,] 0.07475629 0.9756646 0.07475629
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.57286876 2.5998628 3.1414185
#>  [2,] 0.33787707 0.8759482 2.2671282
#>  [3,] 0.08737076 0.8779266 1.9666401
#>  [4,] 4.57326012 0.7512343 0.6558222
#>  [5,] 1.20846656 2.5682691 1.8608606
#>  [6,] 0.03654980 0.0571845 1.7082451
#>  [7,] 0.70636646 0.6391126 1.7077574
#>  [8,] 1.91389337 0.2376174 0.8013366
#>  [9,] 0.01052602 0.1378903 0.5138166
#> [10,] 0.61262104 0.2050553 1.9958753
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 3.3117416 3.3117416 3.3117416
#>  [2,] 0.2876672 0.2876672 0.2876672
#>  [3,] 0.7193227 0.7193227 0.7193227
#>  [4,] 0.1349109 0.1349109 0.1349109
#>  [5,] 1.2165724 1.2165724 1.2165724
#>  [6,] 0.9816001 0.9816001 0.9816001
#>  [7,] 1.2346818 1.2346818 1.2346818
#>  [8,] 0.2989647 0.2989647 0.2989647
#>  [9,] 0.3907461 0.3907461 0.3907461
#> [10,] 0.1301348 0.1301348 0.1301348

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8581193 0.2134336 0.8581193
#>  [2,] 0.6399731 0.4236729 0.2907603
#>  [3,] 0.4194931 0.7709567 0.7709567
#>  [4,] 2.4777434 1.0121171 1.3494673
#>  [5,] 0.2056461 0.2056461 0.2056461
#>  [6,] 0.6234664 0.6234664 0.3042650
#>  [7,] 1.2703919 1.3154818 1.3154818
#>  [8,] 0.4331371 0.4331371 0.7319212
#>  [9,] 2.5900522 2.3472489 1.3160921
#> [10,] 0.1084956 0.1084956 0.1084956
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>               [,1]      [,2]      [,3]
#>  [1,] 0.8962726285 0.2702093 1.2364930
#>  [2,] 0.0008648951 0.4551163 1.3046508
#>  [3,] 0.7161316891 0.1740950 0.7349186
#>  [4,] 0.1552316287 0.1269725 0.0309592
#>  [5,] 2.0152915319 1.5835341 0.5480052
#>  [6,] 0.4729160297 0.2514752 0.1783135
#>  [7,] 0.9637072780 1.6880621 1.7661842
#>  [8,] 1.8244794580 4.4237827 0.2089845
#>  [9,] 0.5700406472 0.3712884 0.6521394
#> [10,] 1.9523436702 0.5723831 0.4489350
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.29603113 1.29603113 1.29603113
#>  [2,] 0.01657664 0.01657664 0.01657664
#>  [3,] 0.36089618 0.36089618 0.36089618
#>  [4,] 1.25743898 1.25743898 1.25743898
#>  [5,] 0.90007510 0.90007510 0.90007510
#>  [6,] 0.11417089 0.11417089 0.11417089
#>  [7,] 1.79241759 1.79241759 1.79241759
#>  [8,] 0.14985060 0.14985060 0.14985060
#>  [9,] 0.95694083 0.95694083 0.95694083
#> [10,] 2.98044067 2.98044067 2.98044067

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1864675 0.1864675 0.1864675
#>  [2,] 1.5442198 1.5442198 1.5442198
#>  [3,] 0.2029502 1.8023260 2.6123207
#>  [4,] 2.1185950 0.5525706 2.2159155
#>  [5,] 1.3186903 0.9230316 0.1409592
#>  [6,] 0.8768026 1.3256219 0.8768026
#>  [7,] 0.1792156 0.3658637 0.1792156
#>  [8,] 1.6055259 1.6055259 1.6055259
#>  [9,] 0.0723159 0.2614596 1.7699896
#> [10,] 1.1936085 1.1936085 1.1936085
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.23517360 1.2443960 0.06384330
#>  [2,] 0.64147302 0.4188675 2.00529027
#>  [3,] 0.84700041 0.4887472 0.82272295
#>  [4,] 0.59074373 0.1520324 0.04735218
#>  [5,] 0.05070894 2.0837545 0.11871012
#>  [6,] 0.63000577 2.8349324 1.04601767
#>  [7,] 3.65110796 0.3759269 1.70066136
#>  [8,] 0.59163398 0.4237969 1.25504789
#>  [9,] 0.76207307 0.2758246 2.49309976
#> [10,] 0.11765470 0.3538048 0.52494261
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1493904 1.1493904 1.1493904
#>  [2,] 0.2644252 0.2644252 0.2644252
#>  [3,] 0.1826504 0.1826504 0.1826504
#>  [4,] 0.4355497 0.4355497 0.4355497
#>  [5,] 0.4437183 0.4437183 0.4437183
#>  [6,] 4.3008913 4.3008913 4.3008913
#>  [7,] 2.4585075 2.4585075 2.4585075
#>  [8,] 0.1758190 0.1758190 0.1758190
#>  [9,] 0.1060762 0.1060762 0.1060762
#> [10,] 2.7388819 2.7388819 2.7388819