Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.98151913 3.98151913 0.28649031
#>  [2,] 0.20032401 0.20032401 2.15753911
#>  [3,] 4.14420231 2.22626033 1.71373221
#>  [4,] 0.34077609 0.34077609 1.40030180
#>  [5,] 1.47170414 0.90763438 1.80312927
#>  [6,] 0.38693996 0.55607767 0.55607767
#>  [7,] 0.05513897 0.05513897 0.05513897
#>  [8,] 1.33406916 3.67797262 1.90830674
#>  [9,] 0.66974124 0.07193109 0.05060339
#> [10,] 0.90354951 0.33698512 2.89408652
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.1916853 0.16154700 1.209788370
#>  [2,] 0.2552193 0.21388418 2.320085987
#>  [3,] 0.1035185 0.16488692 2.621373558
#>  [4,] 0.9565250 2.93474919 0.883084894
#>  [5,] 0.5538188 0.04315224 0.004893210
#>  [6,] 1.6862387 1.30774606 1.149017083
#>  [7,] 0.2623245 0.99257108 0.006892961
#>  [8,] 0.1145372 1.69388110 0.265178871
#>  [9,] 1.8149143 0.27354086 1.207346557
#> [10,] 2.0812861 0.65005307 0.382198316
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5543068 1.5543068 1.5543068
#>  [2,] 0.7681117 0.7681117 0.7681117
#>  [3,] 5.5806606 5.5806606 5.5806606
#>  [4,] 0.6701123 0.6701123 0.6701123
#>  [5,] 1.1236152 1.1236152 1.1236152
#>  [6,] 0.0735110 0.0735110 0.0735110
#>  [7,] 2.3654814 2.3654814 2.3654814
#>  [8,] 0.4634018 0.4634018 0.4634018
#>  [9,] 2.2917632 2.2917632 2.2917632
#> [10,] 0.2422056 0.2422056 0.2422056

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.53478940 0.9006550 0.39336696
#>  [2,] 1.16291537 1.1629154 0.64429871
#>  [3,] 0.58179909 0.5817991 0.58179909
#>  [4,] 0.21555191 0.2155519 0.21555191
#>  [5,] 0.40532337 0.4053234 0.02552139
#>  [6,] 0.27611809 2.7862581 0.51973413
#>  [7,] 0.64033505 0.5743370 0.79416698
#>  [8,] 2.34126674 0.7881950 2.89916018
#>  [9,] 0.01534421 0.8523676 0.01534421
#> [10,] 0.34173674 3.1682014 2.19019232
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>            [,1]       [,2]         [,3]
#>  [1,] 1.8171731 3.12218747 0.6054562039
#>  [2,] 1.9926032 1.27603852 3.0610899138
#>  [3,] 0.9890775 1.43794117 3.7714254010
#>  [4,] 0.4232968 0.27930074 2.8573128471
#>  [5,] 4.6073714 1.01339028 0.0673451343
#>  [6,] 0.3598677 0.26954199 3.4471404102
#>  [7,] 0.8832527 0.09074656 0.0009565914
#>  [8,] 0.4798713 1.45463489 0.3101870973
#>  [9,] 0.6234853 0.49213950 0.5855964211
#> [10,] 0.0465118 0.31602885 0.1447126556
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2029110 0.2029110 0.2029110
#>  [2,] 1.8865424 1.8865424 1.8865424
#>  [3,] 0.9116007 0.9116007 0.9116007
#>  [4,] 1.0177238 1.0177238 1.0177238
#>  [5,] 1.3535239 1.3535239 1.3535239
#>  [6,] 1.2977580 1.2977580 1.2977580
#>  [7,] 0.2313355 0.2313355 0.2313355
#>  [8,] 1.0646906 1.0646906 1.0646906
#>  [9,] 1.3673196 1.3673196 1.3673196
#> [10,] 0.8673000 0.8673000 0.8673000

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1437132 0.1437132 0.1437132
#>  [2,] 2.4966365 2.4966365 0.4711692
#>  [3,] 0.4726011 0.1942682 1.4637203
#>  [4,] 0.0498509 0.4187326 0.4023905
#>  [5,] 1.0875937 1.0875937 1.0875937
#>  [6,] 1.3937728 1.3937728 1.3937728
#>  [7,] 2.1510882 1.7736997 3.1175047
#>  [8,] 0.4295058 0.4295058 0.2092279
#>  [9,] 0.1686601 0.4572693 0.4572693
#> [10,] 4.0427719 4.0427719 4.0427719
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2028122 1.52206434 1.0492978
#>  [2,] 0.3697856 1.09019777 1.1660115
#>  [3,] 3.9517044 0.24453706 1.4109902
#>  [4,] 0.3853083 1.06846156 2.0491656
#>  [5,] 1.1115787 1.31447473 2.6873123
#>  [6,] 0.5674497 0.01892045 2.1062937
#>  [7,] 0.1092407 0.22888299 0.1847669
#>  [8,] 0.4878732 0.52354321 3.1426070
#>  [9,] 0.9541911 1.29623150 3.8515094
#> [10,] 0.7765056 0.45545868 0.1621538
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.31801668 1.31801668 1.31801668
#>  [2,] 2.24544133 2.24544133 2.24544133
#>  [3,] 0.01103203 0.01103203 0.01103203
#>  [4,] 0.70967301 0.70967301 0.70967301
#>  [5,] 1.69844168 1.69844168 1.69844168
#>  [6,] 0.51286817 0.51286817 0.51286817
#>  [7,] 0.47719267 0.47719267 0.47719267
#>  [8,] 1.18490138 1.18490138 1.18490138
#>  [9,] 0.24918033 0.24918033 0.24918033
#> [10,] 1.16712985 1.16712985 1.16712985

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5356288 0.1250970 0.6547737
#>  [2,] 4.6593320 2.6209248 0.9298609
#>  [3,] 1.1724493 1.3061810 0.1341417
#>  [4,] 1.4693141 2.8065917 2.8065917
#>  [5,] 0.2529720 0.4285938 0.4285938
#>  [6,] 0.5446514 0.5446514 0.7761423
#>  [7,] 0.0541898 0.0541898 0.0541898
#>  [8,] 0.3466927 1.2809973 1.2809973
#>  [9,] 1.0220355 1.5038382 1.0748843
#> [10,] 1.7302858 1.7302858 1.7302858
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.40200240 0.20490793 0.5276731
#>  [2,] 2.10686744 1.27135074 2.1192930
#>  [3,] 2.91698530 2.08784978 0.3043190
#>  [4,] 0.27094172 0.49816612 0.6409803
#>  [5,] 2.54976878 1.34791792 0.4302244
#>  [6,] 1.08812455 1.71678905 4.9224605
#>  [7,] 0.47989105 0.66503011 0.2449354
#>  [8,] 0.31549197 0.70725850 0.2561962
#>  [9,] 0.16999505 0.07701758 0.1309447
#> [10,] 0.01284955 0.75004086 0.5625049
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9913466 0.9913466 0.9913466
#>  [2,] 0.3796834 0.3796834 0.3796834
#>  [3,] 0.7904656 0.7904656 0.7904656
#>  [4,] 1.1666779 1.1666779 1.1666779
#>  [5,] 1.4028649 1.4028649 1.4028649
#>  [6,] 0.2162098 0.2162098 0.2162098
#>  [7,] 0.5650077 0.5650077 0.5650077
#>  [8,] 1.3293965 1.3293965 1.3293965
#>  [9,] 0.2530424 0.2530424 0.2530424
#> [10,] 0.4337333 0.4337333 0.4337333