Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>            [,1]       [,2]      [,3]
#>  [1,] 1.8853261 0.39155893 1.8853261
#>  [2,] 2.2174419 1.49434076 1.4184421
#>  [3,] 0.1871281 0.18712811 0.1871281
#>  [4,] 0.1695229 0.68650228 1.2065985
#>  [5,] 0.1761961 0.06088807 0.1179877
#>  [6,] 0.8932845 3.71603788 2.0324845
#>  [7,] 0.0542635 0.05426350 0.1317229
#>  [8,] 0.9169555 0.45959774 0.2087033
#>  [9,] 1.1270539 1.29480320 1.2948032
#> [10,] 0.8778328 0.68656437 1.3979310
rexmo(10, 3, c(3, 0, 0))          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09706466 0.14485111 0.32892811
#>  [2,] 1.24054558 0.08140288 2.08997772
#>  [3,] 0.09645275 1.03758624 0.87849976
#>  [4,] 2.22923492 0.63016280 0.09256857
#>  [5,] 1.35544472 0.77904833 1.17977867
#>  [6,] 2.56390228 0.52566694 0.94222450
#>  [7,] 2.89474000 0.17272443 0.81527127
#>  [8,] 0.55373656 0.64267610 0.27074285
#>  [9,] 0.25644651 1.90197481 1.09490618
#> [10,] 0.69612714 1.13788115 1.79281477
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0010278 1.0010278 1.0010278
#>  [2,] 0.4586296 0.4586296 0.4586296
#>  [3,] 2.2408132 2.2408132 2.2408132
#>  [4,] 0.3457985 0.3457985 0.3457985
#>  [5,] 0.1156603 0.1156603 0.1156603
#>  [6,] 0.1637796 0.1637796 0.1637796
#>  [7,] 0.2170557 0.2170557 0.2170557
#>  [8,] 2.7020384 2.7020384 2.7020384
#>  [9,] 2.7726129 2.7726129 2.7726129
#> [10,] 0.1618028 0.1618028 0.1618028

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 4.2433810 0.17705336 2.60776682
#>  [2,] 0.7154968 0.71549681 0.04589743
#>  [3,] 2.1017426 2.10174263 1.94797737
#>  [4,] 0.6496530 2.85468971 2.87803295
#>  [5,] 3.0757135 0.05464577 0.03191222
#>  [6,] 0.1277643 0.12776431 0.73145109
#>  [7,] 1.9937703 0.22511456 1.99377034
#>  [8,] 0.7751038 1.24151668 1.24151668
#>  [9,] 0.8416479 0.84164795 0.84164795
#> [10,] 2.6844096 2.30354504 2.30354504
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 3.30555070 0.15740354 0.5370576
#>  [2,] 0.68718205 6.71195383 0.8031423
#>  [3,] 1.78822469 0.56985288 0.1368340
#>  [4,] 1.72996665 1.59917195 0.1113620
#>  [5,] 0.05151776 0.14633263 0.1660881
#>  [6,] 0.67300633 1.50655232 0.9973985
#>  [7,] 0.09555423 0.09542277 2.6874981
#>  [8,] 0.02543565 0.05767499 1.7569529
#>  [9,] 0.10008984 2.66966288 0.2017444
#> [10,] 2.79588991 1.46396819 2.6700464
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1012567 0.1012567 0.1012567
#>  [2,] 2.0182928 2.0182928 2.0182928
#>  [3,] 0.2094395 0.2094395 0.2094395
#>  [4,] 1.6071582 1.6071582 1.6071582
#>  [5,] 0.5685241 0.5685241 0.5685241
#>  [6,] 0.6288486 0.6288486 0.6288486
#>  [7,] 0.2722604 0.2722604 0.2722604
#>  [8,] 0.3638800 0.3638800 0.3638800
#>  [9,] 0.6323731 0.6323731 0.6323731
#> [10,] 0.6191974 0.6191974 0.6191974

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.65204812 2.25992758 0.3687901
#>  [2,] 1.52810961 1.52810961 1.5281096
#>  [3,] 0.02516321 0.02516321 0.1748718
#>  [4,] 3.46301601 1.29211789 1.2921179
#>  [5,] 2.13763316 2.13763316 2.1376332
#>  [6,] 0.69356245 0.66709975 0.1233270
#>  [7,] 0.82130632 0.40370032 0.8213063
#>  [8,] 0.79104319 0.38328055 1.8801413
#>  [9,] 1.05153713 0.17402320 1.0515371
#> [10,] 0.52960928 0.52960928 0.7836199
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.27734239 1.0865645 0.43361227
#>  [2,] 0.86622383 0.4025356 0.68730708
#>  [3,] 0.67549174 1.0241808 2.07820500
#>  [4,] 1.04666448 1.2236420 0.01027120
#>  [5,] 0.05831984 3.1521241 1.97576718
#>  [6,] 0.02503510 0.9107448 0.81665988
#>  [7,] 1.59135199 0.8485917 3.10127983
#>  [8,] 2.66702677 0.2878433 2.53333753
#>  [9,] 0.18278066 1.4281345 0.04026456
#> [10,] 0.02244231 2.0619768 0.88439914
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.274051386 0.274051386 0.274051386
#>  [2,] 0.345013502 0.345013502 0.345013502
#>  [3,] 0.500336891 0.500336891 0.500336891
#>  [4,] 0.569651013 0.569651013 0.569651013
#>  [5,] 1.150087910 1.150087910 1.150087910
#>  [6,] 1.281184768 1.281184768 1.281184768
#>  [7,] 0.539753687 0.539753687 0.539753687
#>  [8,] 3.617040200 3.617040200 3.617040200
#>  [9,] 0.006784299 0.006784299 0.006784299
#> [10,] 0.188068704 0.188068704 0.188068704

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20166273 0.46401815 0.15896849
#>  [2,] 1.81397123 2.17849442 0.10912057
#>  [3,] 0.06277225 0.49149848 0.78574549
#>  [4,] 0.76579280 0.27781564 0.76579280
#>  [5,] 2.21766494 3.14611375 0.64317771
#>  [6,] 0.06098396 0.06098396 0.06098396
#>  [7,] 0.83667836 4.06331171 2.29953462
#>  [8,] 0.31011227 0.75346827 0.29521729
#>  [9,] 0.47738891 0.47738891 0.47738891
#> [10,] 0.05243836 0.05243836 0.05243836
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 6.8322384 2.74405077 0.35818192
#>  [2,] 1.9889417 0.73298183 0.21888034
#>  [3,] 0.1024004 0.09838083 0.49520295
#>  [4,] 1.4255158 2.38881561 1.01730401
#>  [5,] 0.2771861 2.62018399 0.01954262
#>  [6,] 0.4612789 2.11666701 0.79761650
#>  [7,] 1.2372968 0.32230984 2.83067517
#>  [8,] 2.5381192 0.92949153 0.58024435
#>  [9,] 0.1156856 0.16798274 0.82636124
#> [10,] 0.9986399 0.25269341 0.14695188
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.25428539 3.25428539 3.25428539
#>  [2,] 1.03932698 1.03932698 1.03932698
#>  [3,] 1.82967539 1.82967539 1.82967539
#>  [4,] 1.24588939 1.24588939 1.24588939
#>  [5,] 2.94204740 2.94204740 2.94204740
#>  [6,] 1.77402161 1.77402161 1.77402161
#>  [7,] 0.47037124 0.47037124 0.47037124
#>  [8,] 0.77425090 0.77425090 0.77425090
#>  [9,] 1.07642052 1.07642052 1.07642052
#> [10,] 0.03779739 0.03779739 0.03779739