Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.92913506 1.8069255 1.9291351
#>  [2,] 1.71493288 0.9483086 0.8347797
#>  [3,] 0.26652757 0.2665276 0.2665276
#>  [4,] 1.29378524 0.8247846 1.5676470
#>  [5,] 0.06807688 4.4898746 0.3193411
#>  [6,] 0.71928103 0.7192810 0.7192810
#>  [7,] 0.42256962 0.5876016 0.3648153
#>  [8,] 5.55852562 1.8941437 0.0952983
#>  [9,] 0.20873072 1.6251575 1.2842703
#> [10,] 0.32616778 0.3261678 0.3261678
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.1132364 0.04967491 2.137506701
#>  [2,] 2.0342936 0.70721607 0.037877396
#>  [3,] 1.0061647 0.08207105 1.064377469
#>  [4,] 0.1520487 0.91584899 0.913382843
#>  [5,] 0.3516360 0.19158868 0.148762164
#>  [6,] 0.1003282 0.77255343 0.007120458
#>  [7,] 0.8343826 2.02162585 1.199120134
#>  [8,] 0.1369607 0.04116284 0.772037362
#>  [9,] 0.7779023 1.18383026 0.220906303
#> [10,] 0.4402176 4.52204859 0.221377636
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9633773 0.9633773 0.9633773
#>  [2,] 1.0686380 1.0686380 1.0686380
#>  [3,] 1.9350450 1.9350450 1.9350450
#>  [4,] 0.1292826 0.1292826 0.1292826
#>  [5,] 1.6900996 1.6900996 1.6900996
#>  [6,] 0.1066214 0.1066214 0.1066214
#>  [7,] 0.1823518 0.1823518 0.1823518
#>  [8,] 0.3986464 0.3986464 0.3986464
#>  [9,] 0.2514606 0.2514606 0.2514606
#> [10,] 0.4998100 0.4998100 0.4998100

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.95709251 3.83314052 1.95709251
#>  [2,] 0.03021946 0.03021946 0.03021946
#>  [3,] 0.69565749 0.69565749 0.69565749
#>  [4,] 1.37280835 0.21805148 0.62181212
#>  [5,] 0.97803013 1.05245651 0.97803013
#>  [6,] 0.13812563 0.13812563 0.13812563
#>  [7,] 0.29628797 0.39035641 0.29628797
#>  [8,] 3.54618052 1.84170122 0.16085048
#>  [9,] 0.48150434 2.61854339 2.61854339
#> [10,] 5.11354026 0.09652845 0.09652845
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]        [,2]      [,3]
#>  [1,] 1.14545920 0.187770490 0.7266347
#>  [2,] 1.27002196 0.143593289 1.2047504
#>  [3,] 2.31340162 1.680786621 0.5405594
#>  [4,] 0.09610259 0.381453433 0.3297624
#>  [5,] 0.70768028 3.729769069 0.2093917
#>  [6,] 0.98673613 0.813217707 0.6580574
#>  [7,] 0.22842019 1.157482327 1.3910553
#>  [8,] 0.24105621 0.001135128 1.8571980
#>  [9,] 0.18071886 1.177246685 0.3700324
#> [10,] 0.18900493 0.326975965 2.1291577
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02852148 0.02852148 0.02852148
#>  [2,] 1.79937172 1.79937172 1.79937172
#>  [3,] 0.67715790 0.67715790 0.67715790
#>  [4,] 0.83836908 0.83836908 0.83836908
#>  [5,] 1.04712868 1.04712868 1.04712868
#>  [6,] 5.87877968 5.87877968 5.87877968
#>  [7,] 0.19536040 0.19536040 0.19536040
#>  [8,] 4.34511976 4.34511976 4.34511976
#>  [9,] 0.11357117 0.11357117 0.11357117
#> [10,] 2.03540166 2.03540166 2.03540166

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41620547 1.19682708 1.31365761
#>  [2,] 0.75141792 1.70183652 0.63025460
#>  [3,] 1.29770715 0.78412112 1.29770715
#>  [4,] 0.80256536 2.77466977 0.80256536
#>  [5,] 0.44156437 1.05221769 0.45204702
#>  [6,] 0.09243823 0.09243823 0.09243823
#>  [7,] 0.11839493 0.11839493 0.11839493
#>  [8,] 0.24339520 0.24339520 0.24339520
#>  [9,] 0.03891272 0.17918494 0.17918494
#> [10,] 0.74241629 0.25615939 0.25615939
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.392058374 0.53860052 0.06461449
#>  [2,] 0.373691924 1.79717438 0.86405343
#>  [3,] 2.051254291 0.66415076 1.79405233
#>  [4,] 0.905056149 1.03388358 0.30066162
#>  [5,] 0.002702792 0.48062153 0.71130788
#>  [6,] 0.037049602 0.06646768 0.75600082
#>  [7,] 2.734852161 0.43010449 0.08468535
#>  [8,] 0.719490581 2.10843281 6.98138290
#>  [9,] 0.533847347 1.16541455 2.56811462
#> [10,] 0.291073096 0.29858941 1.06844016
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3173490 0.3173490 0.3173490
#>  [2,] 0.4894672 0.4894672 0.4894672
#>  [3,] 1.6135888 1.6135888 1.6135888
#>  [4,] 2.1655765 2.1655765 2.1655765
#>  [5,] 1.1899246 1.1899246 1.1899246
#>  [6,] 1.2748108 1.2748108 1.2748108
#>  [7,] 0.3096082 0.3096082 0.3096082
#>  [8,] 0.3559475 0.3559475 0.3559475
#>  [9,] 1.3825644 1.3825644 1.3825644
#> [10,] 0.3641525 0.3641525 0.3641525

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0850906 0.7237470 1.3654251
#>  [2,] 0.4759263 0.4759263 0.4759263
#>  [3,] 0.1524174 0.1524174 0.8272911
#>  [4,] 3.7642125 1.6078803 2.5569532
#>  [5,] 1.0816536 0.2556124 1.8852187
#>  [6,] 0.1328716 0.1328716 0.1328716
#>  [7,] 2.2705582 0.7321872 0.5982103
#>  [8,] 0.3999866 0.3999866 0.3999866
#>  [9,] 0.6969711 2.1469867 0.3213672
#> [10,] 1.0976338 0.3524648 0.1559611
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]        [,2]      [,3]
#>  [1,] 0.4719577 2.887479836 2.9196245
#>  [2,] 0.9600803 4.658286262 0.4718578
#>  [3,] 0.1379359 2.145465765 0.2451174
#>  [4,] 1.5976230 0.619618359 1.0532523
#>  [5,] 1.0350113 1.282775189 2.4298124
#>  [6,] 0.1615191 0.057377705 3.1203480
#>  [7,] 2.0100201 0.006555968 1.9105113
#>  [8,] 0.3142387 0.039998445 0.5702866
#>  [9,] 0.4313515 5.321308732 0.1706717
#> [10,] 0.1314888 0.804589798 1.3292386
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>               [,1]         [,2]         [,3]
#>  [1,] 0.5229525273 0.5229525273 0.5229525273
#>  [2,] 1.4506400618 1.4506400618 1.4506400618
#>  [3,] 0.6251597842 0.6251597842 0.6251597842
#>  [4,] 0.7361193923 0.7361193923 0.7361193923
#>  [5,] 3.6427128667 3.6427128667 3.6427128667
#>  [6,] 0.3206584156 0.3206584156 0.3206584156
#>  [7,] 0.0003428556 0.0003428556 0.0003428556
#>  [8,] 1.6175794057 1.6175794057 1.6175794057
#>  [9,] 0.2278866861 0.2278866861 0.2278866861
#> [10,] 0.4116439046 0.4116439046 0.4116439046