Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3392846 0.06104002 0.3392846
#>  [2,] 0.7825532 6.49064750 1.4230494
#>  [3,] 2.0749023 2.07490230 1.3597937
#>  [4,] 0.2046543 0.20465425 0.2046543
#>  [5,] 2.4651832 0.02745567 3.2860553
#>  [6,] 1.6852503 1.68525034 0.7373269
#>  [7,] 0.5292264 0.52922641 0.5292264
#>  [8,] 0.1982672 0.10651381 0.1065138
#>  [9,] 0.7074811 0.70748110 0.7074811
#> [10,] 2.0140653 1.69447479 1.4168649
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>              [,1]       [,2]       [,3]
#>  [1,] 1.658652566 0.49213597 0.10693677
#>  [2,] 2.113281580 1.25422309 0.68274821
#>  [3,] 1.085895632 3.82233046 0.15492654
#>  [4,] 0.435260223 0.06429493 0.49426619
#>  [5,] 1.507304414 0.93673865 0.54887343
#>  [6,] 1.372368909 2.31462796 0.02176722
#>  [7,] 1.518564613 0.03599016 0.41524983
#>  [8,] 0.003494575 3.31396227 5.85587564
#>  [9,] 0.775419424 0.54639109 0.36726082
#> [10,] 0.511959424 4.74327567 0.86848736
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07260396 1.07260396 1.07260396
#>  [2,] 0.77218601 0.77218601 0.77218601
#>  [3,] 1.15013885 1.15013885 1.15013885
#>  [4,] 1.32951821 1.32951821 1.32951821
#>  [5,] 2.92949053 2.92949053 2.92949053
#>  [6,] 0.27383228 0.27383228 0.27383228
#>  [7,] 0.44179067 0.44179067 0.44179067
#>  [8,] 0.05971237 0.05971237 0.05971237
#>  [9,] 2.97830395 2.97830395 2.97830395
#> [10,] 1.41366007 1.41366007 1.41366007

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.53391585 0.7682946 0.4406974
#>  [2,] 1.69650582 2.8276003 1.1290940
#>  [3,] 0.32169854 1.2918453 4.5812776
#>  [4,] 0.08155124 1.2874671 0.4798332
#>  [5,] 0.61836204 0.6183620 0.6183620
#>  [6,] 1.05385452 1.0538545 1.0538545
#>  [7,] 0.61417276 0.1101919 0.1101919
#>  [8,] 1.59391971 4.0155754 4.6533302
#>  [9,] 0.72631534 0.7263153 0.6721811
#> [10,] 0.30614061 0.2169383 0.2169383
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.79105422 0.17179500 0.6753164
#>  [2,] 0.13365477 1.29432716 4.6079542
#>  [3,] 0.01712518 0.95340787 0.1253108
#>  [4,] 0.82406357 1.13766848 0.7194237
#>  [5,] 0.42732343 0.23020793 0.3717771
#>  [6,] 2.10116292 0.86709502 0.7608435
#>  [7,] 1.71778129 2.27339481 2.0786914
#>  [8,] 1.90548036 0.01857295 1.9624886
#>  [9,] 3.98620731 0.38646367 0.5312779
#> [10,] 0.18236563 0.68162876 0.8260901
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1750215 0.1750215 0.1750215
#>  [2,] 0.3126704 0.3126704 0.3126704
#>  [3,] 1.4132032 1.4132032 1.4132032
#>  [4,] 2.8119707 2.8119707 2.8119707
#>  [5,] 0.5563832 0.5563832 0.5563832
#>  [6,] 1.2541737 1.2541737 1.2541737
#>  [7,] 1.3579134 1.3579134 1.3579134
#>  [8,] 0.9044375 0.9044375 0.9044375
#>  [9,] 1.0790248 1.0790248 1.0790248
#> [10,] 0.3327255 0.3327255 0.3327255

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>            [,1]       [,2]        [,3]
#>  [1,] 2.2254319 0.60363130 0.200079297
#>  [2,] 2.6748769 1.11968364 2.674876935
#>  [3,] 0.3283678 1.72464821 1.336150936
#>  [4,] 0.2544494 3.70101807 1.037528376
#>  [5,] 2.9348062 5.05421240 0.130718489
#>  [6,] 3.3039580 1.32175265 0.805178079
#>  [7,] 2.1553173 0.02080368 0.008189154
#>  [8,] 1.0005604 0.34064852 2.763737650
#>  [9,] 0.1535855 5.72952817 0.153585539
#> [10,] 0.9236616 0.10241008 0.072723296
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.84511067 1.1739887 0.05328529
#>  [2,] 0.05740727 0.4916203 1.05159211
#>  [3,] 1.51374702 0.2657493 1.18599289
#>  [4,] 0.02956646 2.6861917 0.59470263
#>  [5,] 2.43374544 0.6018099 1.20236884
#>  [6,] 0.66751184 2.6658085 0.36609681
#>  [7,] 0.18123592 0.7586609 0.14052013
#>  [8,] 1.53884622 2.5398502 0.42833178
#>  [9,] 0.07601225 1.4069822 0.37155355
#> [10,] 0.10488380 0.2526422 0.11677134
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.88015895 0.88015895 0.88015895
#>  [2,] 0.96087402 0.96087402 0.96087402
#>  [3,] 0.40671017 0.40671017 0.40671017
#>  [4,] 1.78314616 1.78314616 1.78314616
#>  [5,] 1.11577533 1.11577533 1.11577533
#>  [6,] 0.06242764 0.06242764 0.06242764
#>  [7,] 0.63968355 0.63968355 0.63968355
#>  [8,] 0.63928969 0.63928969 0.63928969
#>  [9,] 0.32909766 0.32909766 0.32909766
#> [10,] 0.59202451 0.59202451 0.59202451

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.7911637 0.4125214 1.11364306
#>  [2,] 1.4468109 0.6566485 1.44681089
#>  [3,] 1.8839631 1.2631306 2.85052417
#>  [4,] 0.4442270 0.7587278 0.44422703
#>  [5,] 0.9116316 0.4852580 1.14467599
#>  [6,] 0.8958303 2.3008107 2.30081068
#>  [7,] 6.5351771 0.8746499 0.08355910
#>  [8,] 2.5062564 2.0150704 2.01507040
#>  [9,] 0.6982791 1.1963593 0.69827913
#> [10,] 0.3981631 1.1321795 0.08464002
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.71008680 0.8550921 0.96623646
#>  [2,] 0.61631625 2.6568681 0.95799373
#>  [3,] 0.10236016 1.0499270 4.28680943
#>  [4,] 0.44710070 0.4347386 2.06863285
#>  [5,] 1.37973422 1.3469453 1.91840400
#>  [6,] 1.20972511 1.1369966 0.49069532
#>  [7,] 0.01683796 0.1565956 5.78387962
#>  [8,] 0.61439683 0.6546543 1.06987732
#>  [9,] 4.22728779 1.8793435 6.44940551
#> [10,] 0.09414283 1.6268073 0.03365184
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.22567648 0.22567648 0.22567648
#>  [2,] 0.04357378 0.04357378 0.04357378
#>  [3,] 0.32536691 0.32536691 0.32536691
#>  [4,] 0.11002354 0.11002354 0.11002354
#>  [5,] 0.18070123 0.18070123 0.18070123
#>  [6,] 2.50519553 2.50519553 2.50519553
#>  [7,] 1.60444885 1.60444885 1.60444885
#>  [8,] 0.33577686 0.33577686 0.33577686
#>  [9,] 0.01749702 0.01749702 0.01749702
#> [10,] 1.48516066 1.48516066 1.48516066