Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>             [,1]      [,2]      [,3]
#>  [1,] 3.71694267 1.9686437 2.5087213
#>  [2,] 0.41634640 0.2906725 0.6407367
#>  [3,] 0.20179523 1.0935207 0.7653752
#>  [4,] 1.35317313 1.3531731 1.3531731
#>  [5,] 1.70473123 0.8157264 0.4095571
#>  [6,] 1.58573021 1.5857302 0.5370731
#>  [7,] 0.28602609 0.2860261 0.2860261
#>  [8,] 0.20457822 0.4879310 0.2045782
#>  [9,] 0.07525108 1.1406774 0.3790559
#> [10,] 0.36084213 0.3608421 0.3608421
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>            [,1]        [,2]      [,3]
#>  [1,] 1.4916099 0.002242199 0.5363157
#>  [2,] 0.4796558 1.505709123 0.4181797
#>  [3,] 2.2801178 2.474058997 2.2712915
#>  [4,] 0.4241598 0.278050408 1.2225842
#>  [5,] 0.2993507 5.663360587 0.0703431
#>  [6,] 0.4212098 0.003681975 0.7905407
#>  [7,] 0.1429447 1.091104210 1.4482503
#>  [8,] 8.0720292 0.577717560 1.2355541
#>  [9,] 1.1723097 0.953341266 0.1084903
#> [10,] 0.3530467 0.072831895 0.2826922
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42919588 0.42919588 0.42919588
#>  [2,] 6.37015711 6.37015711 6.37015711
#>  [3,] 0.19538213 0.19538213 0.19538213
#>  [4,] 0.13870721 0.13870721 0.13870721
#>  [5,] 1.55338340 1.55338340 1.55338340
#>  [6,] 0.09768018 0.09768018 0.09768018
#>  [7,] 2.32903925 2.32903925 2.32903925
#>  [8,] 0.39125070 0.39125070 0.39125070
#>  [9,] 0.08161856 0.08161856 0.08161856
#> [10,] 0.83966962 0.83966962 0.83966962

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05727960 0.76347424 1.35785432
#>  [2,] 1.77425957 2.09638786 0.36403695
#>  [3,] 0.07739163 0.07739163 0.07739163
#>  [4,] 3.45410271 3.63212610 3.48668321
#>  [5,] 0.30593266 0.30593266 0.30593266
#>  [6,] 0.92796015 0.92796015 0.92796015
#>  [7,] 2.62392469 0.81696723 0.54434919
#>  [8,] 0.17364482 1.31984278 1.31984278
#>  [9,] 0.27055848 0.27055848 0.27055848
#> [10,] 0.27617381 0.19586459 0.27617381
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>              [,1]       [,2]      [,3]
#>  [1,] 1.435663723 2.92694185 0.7113739
#>  [2,] 0.001191443 0.82405906 0.1304247
#>  [3,] 2.058032746 0.83305283 1.5238540
#>  [4,] 1.487589608 0.09009481 0.8792486
#>  [5,] 0.549662349 2.10544699 0.5324749
#>  [6,] 0.459098396 0.16203684 0.1693327
#>  [7,] 0.127117567 0.41051683 1.5918131
#>  [8,] 0.019494521 0.91340954 3.1424910
#>  [9,] 0.834080877 4.24016564 4.8735408
#> [10,] 0.160671876 0.74283151 0.1289182
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02110038 0.02110038 0.02110038
#>  [2,] 1.73442830 1.73442830 1.73442830
#>  [3,] 0.22772223 0.22772223 0.22772223
#>  [4,] 0.07016357 0.07016357 0.07016357
#>  [5,] 0.26750661 0.26750661 0.26750661
#>  [6,] 0.96636804 0.96636804 0.96636804
#>  [7,] 1.81160382 1.81160382 1.81160382
#>  [8,] 0.67729726 0.67729726 0.67729726
#>  [9,] 2.17567214 2.17567214 2.17567214
#> [10,] 2.40614719 2.40614719 2.40614719

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1215802 0.12158024 0.12158024
#>  [2,] 0.5968824 0.04414442 1.83420680
#>  [3,] 0.3509615 1.21818168 0.04721186
#>  [4,] 0.3884908 0.38849082 0.38849082
#>  [5,] 1.5488418 0.39249897 1.54884177
#>  [6,] 1.2431209 1.12936921 1.12936921
#>  [7,] 2.1323315 0.75664862 0.75664862
#>  [8,] 0.0628163 1.56753757 1.42972205
#>  [9,] 0.7703106 1.05347221 0.77031058
#> [10,] 1.3301629 2.05807776 0.44743656
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.106966626 0.1464768 1.5066436
#>  [2,] 1.681433598 1.5091183 1.7836933
#>  [3,] 0.665684796 0.1741431 0.5804785
#>  [4,] 2.023251006 0.2139639 1.9598998
#>  [5,] 2.612144627 0.5635894 3.3750703
#>  [6,] 0.778321268 0.3476337 0.3360992
#>  [7,] 0.428441828 2.0547191 0.3040250
#>  [8,] 0.008754794 2.8307979 1.5337556
#>  [9,] 0.901226976 0.1731149 0.6213585
#> [10,] 1.889853040 4.0497861 0.1625730
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9097049 0.9097049 0.9097049
#>  [2,] 0.1420463 0.1420463 0.1420463
#>  [3,] 0.3283024 0.3283024 0.3283024
#>  [4,] 0.5212641 0.5212641 0.5212641
#>  [5,] 2.0749054 2.0749054 2.0749054
#>  [6,] 0.2965431 0.2965431 0.2965431
#>  [7,] 0.7743797 0.7743797 0.7743797
#>  [8,] 1.3598281 1.3598281 1.3598281
#>  [9,] 0.3977140 0.3977140 0.3977140
#> [10,] 0.2051140 0.2051140 0.2051140

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5406762 1.4699864 1.1478033
#>  [2,] 0.1379454 0.1379454 1.5579339
#>  [3,] 0.7108135 0.7108135 0.7108135
#>  [4,] 3.3769834 1.9624090 0.3412477
#>  [5,] 2.6701219 3.4743822 0.2078564
#>  [6,] 0.5170115 0.3167925 0.5170115
#>  [7,] 0.8017059 0.5810908 0.8017059
#>  [8,] 0.9599623 0.9599623 0.9599623
#>  [9,] 1.1456828 0.5610817 1.4523435
#> [10,] 0.5621723 0.5862144 1.3203438
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6075910 2.73126382 2.52675175
#>  [2,] 1.1380242 0.49286172 0.26839810
#>  [3,] 0.9253590 1.18188380 1.60971769
#>  [4,] 0.2981943 0.02708347 0.48294578
#>  [5,] 0.8855871 2.30535570 3.01125051
#>  [6,] 0.2555675 2.28851257 0.05232346
#>  [7,] 1.4471666 0.80615898 0.80443529
#>  [8,] 0.1311776 1.10463096 0.36381391
#>  [9,] 0.1243688 3.24607427 0.70324312
#> [10,] 0.8306983 1.46548356 1.30415764
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13911786 0.13911786 0.13911786
#>  [2,] 0.64969063 0.64969063 0.64969063
#>  [3,] 1.24974339 1.24974339 1.24974339
#>  [4,] 0.29756146 0.29756146 0.29756146
#>  [5,] 0.35642151 0.35642151 0.35642151
#>  [6,] 1.63244487 1.63244487 1.63244487
#>  [7,] 1.53688190 1.53688190 1.53688190
#>  [8,] 0.04347251 0.04347251 0.04347251
#>  [9,] 1.04657380 1.04657380 1.04657380
#> [10,] 0.06034519 0.06034519 0.06034519