Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationship of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.8518057 2.69752465 1.24600790
#>  [2,] 1.1424554 0.46377095 4.79433181
#>  [3,] 3.3730995 0.42725454 0.42725454
#>  [4,] 0.6275251 0.45091608 0.23802097
#>  [5,] 0.4040373 0.40403725 0.40403725
#>  [6,] 0.2871488 0.28714884 0.28714884
#>  [7,] 0.7268867 1.29787248 0.42705887
#>  [8,] 0.4390868 1.75961458 1.75961458
#>  [9,] 1.2689907 0.70877913 0.08665049
#> [10,] 0.2409586 0.07579288 0.10754536
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.29797084 1.1540098 0.50209952
#>  [2,] 1.50537394 0.1625223 2.49044136
#>  [3,] 1.06355659 0.6521616 0.21265140
#>  [4,] 1.51215090 0.9140000 0.52673191
#>  [5,] 1.90913655 1.8657892 0.64785953
#>  [6,] 0.09689082 0.5464291 0.02362812
#>  [7,] 0.74955424 0.1921662 0.02445247
#>  [8,] 0.16401927 0.5298247 0.58367046
#>  [9,] 1.23104450 0.1592426 0.88922662
#> [10,] 3.05302021 0.5791366 0.65369718
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.48945653 0.48945653 0.48945653
#>  [2,] 0.03888861 0.03888861 0.03888861
#>  [3,] 0.67296160 0.67296160 0.67296160
#>  [4,] 0.47966900 0.47966900 0.47966900
#>  [5,] 0.06543486 0.06543486 0.06543486
#>  [6,] 1.67231264 1.67231264 1.67231264
#>  [7,] 0.12275205 0.12275205 0.12275205
#>  [8,] 8.36015793 8.36015793 8.36015793
#>  [9,] 1.09456012 1.09456012 1.09456012
#> [10,] 0.40472572 0.40472572 0.40472572

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]        [,2]        [,3]
#>  [1,] 3.23303033 0.683787753 3.233030332
#>  [2,] 0.46938865 4.350326587 0.969288718
#>  [3,] 0.12216011 1.181826090 0.165810033
#>  [4,] 0.35601924 0.332747939 0.332747939
#>  [5,] 0.05134016 0.051340158 0.245299710
#>  [6,] 0.45864787 0.006679882 0.006679882
#>  [7,] 1.95287114 1.952871136 1.124474563
#>  [8,] 0.93702536 0.409428790 0.409428790
#>  [9,] 1.48773027 0.552071691 0.552071691
#> [10,] 0.55199440 0.798962415 1.809480840
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.07136840 2.0326225 0.22369360
#>  [2,] 0.09522295 0.6005192 1.03198834
#>  [3,] 0.11317581 0.2782549 0.21388592
#>  [4,] 0.77787066 0.3380565 0.03958457
#>  [5,] 1.20365075 2.1294258 2.67431402
#>  [6,] 0.53433583 0.9777496 0.01885765
#>  [7,] 1.17258850 1.9769766 1.90985707
#>  [8,] 1.17886423 0.5287184 0.21141290
#>  [9,] 0.02031109 0.7035882 5.41533468
#> [10,] 0.49048576 0.7421496 0.89325737
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.92976614 1.92976614 1.92976614
#>  [2,] 0.05899967 0.05899967 0.05899967
#>  [3,] 0.35296495 0.35296495 0.35296495
#>  [4,] 0.69994433 0.69994433 0.69994433
#>  [5,] 0.35323145 0.35323145 0.35323145
#>  [6,] 2.06131684 2.06131684 2.06131684
#>  [7,] 0.41889226 0.41889226 0.41889226
#>  [8,] 2.16372511 2.16372511 2.16372511
#>  [9,] 2.99181854 2.99181854 2.99181854
#> [10,] 4.25865721 4.25865721 4.25865721

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 3.02226868 4.88329587 2.5583745
#>  [2,] 1.18097617 1.18097617 1.1809762
#>  [3,] 0.37784065 1.23650421 0.3778407
#>  [4,] 0.16109898 0.43222911 0.1610990
#>  [5,] 2.36547974 0.64315516 0.0254385
#>  [6,] 0.30608399 0.99131430 0.7831267
#>  [7,] 0.18739572 1.95066391 2.7581660
#>  [8,] 0.01338381 0.01338381 2.5622185
#>  [9,] 2.59852886 0.41330959 1.5710022
#> [10,] 0.24009766 0.24009766 0.2400977
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 2.20694183 1.5936085 1.8621739
#>  [2,] 2.93574244 0.9966557 3.0758456
#>  [3,] 0.22074159 3.1822048 1.5141070
#>  [4,] 0.02523734 1.9784024 0.1353713
#>  [5,] 1.50589188 0.4058354 0.7675601
#>  [6,] 1.30707068 2.4477818 1.4863064
#>  [7,] 0.54165355 0.7324884 2.1702710
#>  [8,] 0.39807508 1.6242970 0.3689203
#>  [9,] 0.03569522 1.8312722 0.3831368
#> [10,] 0.30841740 0.3370371 0.1503749
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02567805 0.02567805 0.02567805
#>  [2,] 2.07056874 2.07056874 2.07056874
#>  [3,] 1.55676219 1.55676219 1.55676219
#>  [4,] 1.58226821 1.58226821 1.58226821
#>  [5,] 0.52927782 0.52927782 0.52927782
#>  [6,] 0.72329597 0.72329597 0.72329597
#>  [7,] 0.93640620 0.93640620 0.93640620
#>  [8,] 2.30797384 2.30797384 2.30797384
#>  [9,] 2.22311733 2.22311733 2.22311733
#> [10,] 1.86363405 1.86363405 1.86363405

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4682833 0.3868952 0.1915801
#>  [2,] 0.6459272 0.1860231 1.1105805
#>  [3,] 0.4347553 0.4347553 0.4347553
#>  [4,] 0.1808212 0.1808212 0.1808212
#>  [5,] 3.1195002 3.9766159 0.2956398
#>  [6,] 0.4200357 1.4225589 1.4225589
#>  [7,] 0.3019243 0.1520529 0.1135006
#>  [8,] 0.1759145 1.0150249 0.1140975
#>  [9,] 4.9710806 0.1265801 1.3415307
#> [10,] 0.3151548 1.3142199 2.6760071
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 3.1707907 0.01286845 0.77063833
#>  [2,] 0.1989198 0.16078325 1.44209137
#>  [3,] 0.4147199 0.06803771 1.00844389
#>  [4,] 0.7721740 0.47778337 2.55293514
#>  [5,] 0.1602641 0.32666530 0.02323520
#>  [6,] 2.3421425 0.67316947 2.76164708
#>  [7,] 0.3505231 2.59782529 0.02234072
#>  [8,] 1.6223380 0.09521787 0.30849746
#>  [9,] 0.2225138 1.05237540 0.24857515
#> [10,] 1.2762703 1.03594487 2.24308881
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2529480 1.2529480 1.2529480
#>  [2,] 1.4488553 1.4488553 1.4488553
#>  [3,] 0.8700250 0.8700250 0.8700250
#>  [4,] 2.0545904 2.0545904 2.0545904
#>  [5,] 0.9374743 0.9374743 0.9374743
#>  [6,] 0.4587449 0.4587449 0.4587449
#>  [7,] 5.7345752 5.7345752 5.7345752
#>  [8,] 0.8441013 0.8441013 0.8441013
#>  [9,] 0.3473254 0.3473254 0.3473254
#> [10,] 1.2348789 1.2348789 1.2348789