Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>            [,1]      [,2]        [,3]
#>  [1,] 3.2253491 0.4289510 0.428950951
#>  [2,] 0.1427539 0.1427539 0.319032310
#>  [3,] 2.6549466 1.0697135 1.069713489
#>  [4,] 1.5730521 0.3595915 0.783353069
#>  [5,] 0.3303690 0.3303690 0.330368989
#>  [6,] 0.6568684 0.6568684 0.008385336
#>  [7,] 0.5549872 0.5549872 0.554987233
#>  [8,] 0.1656729 0.2428690 0.165672863
#>  [9,] 0.4204640 0.3171919 3.404817606
#> [10,] 4.5771054 0.7466135 0.746613476
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6407338 0.6145625 0.32248696
#>  [2,] 1.0854290 0.5240119 1.66635363
#>  [3,] 0.4937559 2.4307556 2.12289138
#>  [4,] 1.4477412 0.4653889 1.67592801
#>  [5,] 0.2740176 0.6207507 0.04811315
#>  [6,] 0.1246937 0.8420922 1.39852588
#>  [7,] 2.1327952 2.4765455 1.44707595
#>  [8,] 0.3445885 0.4300091 2.41075785
#>  [9,] 2.2681369 0.9270810 2.24261572
#> [10,] 2.5659245 1.1548023 1.79661008
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2374693 0.2374693 0.2374693
#>  [2,] 0.7256119 0.7256119 0.7256119
#>  [3,] 1.8399470 1.8399470 1.8399470
#>  [4,] 0.9487056 0.9487056 0.9487056
#>  [5,] 0.2315989 0.2315989 0.2315989
#>  [6,] 0.5673419 0.5673419 0.5673419
#>  [7,] 0.3967534 0.3967534 0.3967534
#>  [8,] 3.6308279 3.6308279 3.6308279
#>  [9,] 0.8335561 0.8335561 0.8335561
#> [10,] 0.1374901 0.1374901 0.1374901

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.44792136 0.4479214 0.64168147
#>  [2,] 0.19015163 0.9490215 0.97946092
#>  [3,] 0.20927168 0.1522278 0.15222776
#>  [4,] 0.33519485 0.8488006 0.03127206
#>  [5,] 0.28893119 1.8334008 1.49556333
#>  [6,] 1.08788524 2.4141977 1.08788524
#>  [7,] 0.01483725 0.2708297 1.26079045
#>  [8,] 1.48137225 0.6669789 0.32693697
#>  [9,] 1.88948318 0.9584833 1.09945526
#> [10,] 0.32903610 1.6278717 0.32903610
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.23178712 0.26958583 2.9447504
#>  [2,] 2.32502076 0.13961208 0.1256585
#>  [3,] 0.78178542 0.47833558 0.2251530
#>  [4,] 0.02467088 2.64118327 2.0148978
#>  [5,] 3.42647164 0.08259084 2.0639144
#>  [6,] 6.10947655 0.05333718 1.1735965
#>  [7,] 3.62186843 0.62350708 1.0287572
#>  [8,] 0.51337204 1.13801175 0.2766460
#>  [9,] 2.55816718 0.05929877 0.8296910
#> [10,] 0.21711350 0.07617772 1.3020993
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>               [,1]         [,2]         [,3]
#>  [1,] 0.1736796992 0.1736796992 0.1736796992
#>  [2,] 0.6168529377 0.6168529377 0.6168529377
#>  [3,] 0.6565959007 0.6565959007 0.6565959007
#>  [4,] 0.5660284078 0.5660284078 0.5660284078
#>  [5,] 1.1638390934 1.1638390934 1.1638390934
#>  [6,] 0.5574220549 0.5574220549 0.5574220549
#>  [7,] 0.3572180467 0.3572180467 0.3572180467
#>  [8,] 0.3389851972 0.3389851972 0.3389851972
#>  [9,] 2.6556994934 2.6556994934 2.6556994934
#> [10,] 0.0009612325 0.0009612325 0.0009612325

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.99413790 0.3292037 0.21724524
#>  [2,] 0.07846185 0.6545333 0.73200226
#>  [3,] 0.22544105 2.8126399 2.22246591
#>  [4,] 0.57728601 0.5772860 0.57728601
#>  [5,] 0.50349008 0.5034901 0.65730558
#>  [6,] 0.03061716 0.5937856 0.59378561
#>  [7,] 0.53783234 0.5378323 0.53783234
#>  [8,] 0.56510939 0.5651094 0.77424220
#>  [9,] 0.33432363 0.3196054 1.24026301
#> [10,] 0.97652481 0.9765248 0.02403196
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.25278484 1.41717928 3.1822733
#>  [2,] 0.86605491 0.96124027 0.5888036
#>  [3,] 1.35507471 0.23295591 0.6440757
#>  [4,] 0.57605500 0.08764983 0.4519813
#>  [5,] 0.71685023 3.54358406 0.8543288
#>  [6,] 0.83425064 0.62351152 0.4041736
#>  [7,] 0.05867163 0.46377172 1.0022389
#>  [8,] 0.13186080 0.40432675 3.3091519
#>  [9,] 1.81680185 1.57770037 0.5924220
#> [10,] 1.51696147 0.91730006 0.1456673
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3981328 0.3981328 0.3981328
#>  [2,] 0.7190196 0.7190196 0.7190196
#>  [3,] 1.3859627 1.3859627 1.3859627
#>  [4,] 0.3561279 0.3561279 0.3561279
#>  [5,] 2.1839271 2.1839271 2.1839271
#>  [6,] 1.4245752 1.4245752 1.4245752
#>  [7,] 2.2521291 2.2521291 2.2521291
#>  [8,] 0.5317164 0.5317164 0.5317164
#>  [9,] 0.7842239 0.7842239 0.7842239
#> [10,] 3.3666041 3.3666041 3.3666041

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.58703176 0.58703176 1.3665164
#>  [2,] 0.51653509 0.51653509 2.8554397
#>  [3,] 2.31218315 1.23248846 1.4709725
#>  [4,] 0.98023335 0.98023335 0.9802334
#>  [5,] 2.24867758 1.60796106 0.1632973
#>  [6,] 1.53041930 1.23222374 1.4004872
#>  [7,] 1.09990243 1.50803552 1.6324995
#>  [8,] 0.86358382 0.86358382 1.1473132
#>  [9,] 0.26518853 0.26518853 0.1551813
#> [10,] 0.01359022 0.04898632 1.3877290
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]      [,2]        [,3]
#>  [1,] 0.21527628 3.0166147 1.441939240
#>  [2,] 1.30257851 0.1704381 0.405891311
#>  [3,] 1.42426490 0.9210000 0.453512874
#>  [4,] 0.07853996 0.8765171 1.489944754
#>  [5,] 0.59699140 0.1193266 3.200716562
#>  [6,] 0.13227544 1.9328792 0.171858830
#>  [7,] 1.11972524 0.4202763 0.264442961
#>  [8,] 0.49652596 0.8199659 0.003956667
#>  [9,] 0.19930360 0.1092570 0.374126201
#> [10,] 1.11604836 0.1458004 3.572782796
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.23829656 1.23829656 1.23829656
#>  [2,] 0.88437541 0.88437541 0.88437541
#>  [3,] 0.52183863 0.52183863 0.52183863
#>  [4,] 1.24165420 1.24165420 1.24165420
#>  [5,] 0.24517918 0.24517918 0.24517918
#>  [6,] 0.06435413 0.06435413 0.06435413
#>  [7,] 1.45264988 1.45264988 1.45264988
#>  [8,] 0.12470222 0.12470222 0.12470222
#>  [9,] 2.12498314 2.12498314 2.12498314
#> [10,] 2.64537200 2.64537200 2.64537200