Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensitiesex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>              [,1]       [,2]        [,3]
#>  [1,] 1.383503517 0.60180632 0.191575947
#>  [2,] 0.027247198 0.02724720 0.027247198
#>  [3,] 0.086293006 0.08629301 0.086293006
#>  [4,] 2.345698513 2.09064113 3.391671041
#>  [5,] 1.498910657 2.02842390 1.471980496
#>  [6,] 2.457142506 0.50450064 0.024660047
#>  [7,] 1.511019025 1.28584465 0.040004208
#>  [8,] 1.791038563 1.08512799 1.268829704
#>  [9,] 0.005428076 2.08816032 0.005428076
#> [10,] 0.647784152 0.64778415 0.647784152
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>            [,1]       [,2]       [,3]
#>  [1,] 0.9222535 0.32760929 1.75431080
#>  [2,] 0.3524749 0.28882652 0.28460115
#>  [3,] 0.9765666 0.43799549 0.04922885
#>  [4,] 0.1885802 0.15932468 0.21924122
#>  [5,] 0.5139308 1.51444801 0.33418605
#>  [6,] 2.2137165 0.37562382 3.00475042
#>  [7,] 0.0368538 0.07759846 0.11286265
#>  [8,] 1.4610404 1.27712736 0.94102096
#>  [9,] 1.1790577 0.49080981 1.90035180
#> [10,] 0.4483817 0.92706724 0.06641903
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4854261 0.4854261 0.4854261
#>  [2,] 0.2040681 0.2040681 0.2040681
#>  [3,] 2.6377964 2.6377964 2.6377964
#>  [4,] 0.6225506 0.6225506 0.6225506
#>  [5,] 1.1401959 1.1401959 1.1401959
#>  [6,] 0.5356232 0.5356232 0.5356232
#>  [7,] 1.9429374 1.9429374 1.9429374
#>  [8,] 0.6171887 0.6171887 0.6171887
#>  [9,] 4.6706046 4.6706046 4.6706046
#> [10,] 1.4490139 1.4490139 1.4490139

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.09406408 0.8991729 0.2504970
#>  [2,] 1.61904206 3.3368240 1.1179423
#>  [3,] 1.50000624 0.6261473 2.9007018
#>  [4,] 0.12099220 0.1209922 0.1209922
#>  [5,] 5.93202261 0.8569749 3.7266862
#>  [6,] 0.22160366 0.5353038 0.5353038
#>  [7,] 1.91709167 0.8738789 1.0546752
#>  [8,] 0.27165920 0.2716592 0.2716592
#>  [9,] 0.00828246 0.3978028 2.5542116
#> [10,] 2.58544389 0.5789590 2.5854439
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27484991 0.64122598 0.54268712
#>  [2,] 0.58569742 0.49945400 0.05891535
#>  [3,] 0.34853014 1.00146253 1.48885047
#>  [4,] 0.09110625 0.27255583 0.32626665
#>  [5,] 0.08771201 0.04585194 1.03640128
#>  [6,] 0.19242590 0.75705953 1.63641362
#>  [7,] 0.10961591 2.14606315 0.73103456
#>  [8,] 1.15911802 1.50643868 0.03157947
#>  [9,] 1.71491665 1.53860728 0.91283610
#> [10,] 0.74813228 0.34899933 0.31448779
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5083758 0.5083758 0.5083758
#>  [2,] 0.5965655 0.5965655 0.5965655
#>  [3,] 0.9751626 0.9751626 0.9751626
#>  [4,] 0.7888206 0.7888206 0.7888206
#>  [5,] 1.0064859 1.0064859 1.0064859
#>  [6,] 0.8998450 0.8998450 0.8998450
#>  [7,] 1.2949952 1.2949952 1.2949952
#>  [8,] 1.4033481 1.4033481 1.4033481
#>  [9,] 1.4985868 1.4985868 1.4985868
#> [10,] 1.3376552 1.3376552 1.3376552

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3246625 0.48554452 0.4855445
#>  [2,] 0.7228795 2.04705241 0.1034208
#>  [3,] 1.1995743 0.09028077 0.2759704
#>  [4,] 1.3081483 0.25893260 0.3426837
#>  [5,] 1.0597006 1.53011768 2.2670758
#>  [6,] 0.7574209 2.30871125 0.7574209
#>  [7,] 0.9641716 0.30022582 0.2806477
#>  [8,] 1.1457931 0.79116854 0.2198041
#>  [9,] 0.3103788 0.31037881 0.3103788
#> [10,] 1.3484734 0.52082201 0.7739818
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9951233 0.7838205 0.64947768
#>  [2,] 0.1286544 2.7581000 0.76217723
#>  [3,] 2.6631462 1.6100737 0.01738475
#>  [4,] 0.8477663 1.1195150 1.79399728
#>  [5,] 0.4708387 0.3499997 1.12677200
#>  [6,] 0.4858598 0.8973256 2.01860276
#>  [7,] 0.8915492 1.1394588 0.39077047
#>  [8,] 0.9113065 3.0416784 2.35791081
#>  [9,] 0.3172728 1.9873583 0.66962890
#> [10,] 1.6634256 1.5863532 0.04369672
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.83811517 1.83811517 1.83811517
#>  [2,] 0.53221392 0.53221392 0.53221392
#>  [3,] 0.17306714 0.17306714 0.17306714
#>  [4,] 0.08773078 0.08773078 0.08773078
#>  [5,] 3.31320622 3.31320622 3.31320622
#>  [6,] 0.03818029 0.03818029 0.03818029
#>  [7,] 0.34510938 0.34510938 0.34510938
#>  [8,] 0.21911336 0.21911336 0.21911336
#>  [9,] 0.53699937 0.53699937 0.53699937
#> [10,] 1.23773301 1.23773301 1.23773301

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6737340 0.8625860 0.5869854
#>  [2,] 0.9824350 1.0522739 0.9866861
#>  [3,] 1.2535021 1.2113267 1.2113267
#>  [4,] 0.9008560 0.3464760 1.2542278
#>  [5,] 0.4434171 0.4434171 0.4434171
#>  [6,] 1.1883290 1.2123008 0.8132617
#>  [7,] 0.1109050 0.1109050 1.5567662
#>  [8,] 0.1987594 0.2389307 0.2389307
#>  [9,] 0.8371992 2.0369367 4.0295941
#> [10,] 0.4512510 1.7427031 2.1246636
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09671183 0.34335969 2.85701573
#>  [2,] 0.36450807 0.90305832 0.10382260
#>  [3,] 1.75357050 0.33210694 0.83442333
#>  [4,] 0.30066951 0.12524580 0.78872033
#>  [5,] 1.40578369 0.11703499 0.19845697
#>  [6,] 0.25665801 3.85351657 0.03781236
#>  [7,] 1.16200170 0.09491306 0.04053177
#>  [8,] 0.03097361 2.11624024 0.05815956
#>  [9,] 0.07693629 0.46133288 0.37952145
#> [10,] 2.84618855 0.06980875 3.29679096
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2301097 0.2301097 0.2301097
#>  [2,] 0.6509635 0.6509635 0.6509635
#>  [3,] 1.1942492 1.1942492 1.1942492
#>  [4,] 1.6741771 1.6741771 1.6741771
#>  [5,] 0.4938386 0.4938386 0.4938386
#>  [6,] 0.1515190 0.1515190 0.1515190
#>  [7,] 0.7120333 0.7120333 0.7120333
#>  [8,] 0.3794284 0.3794284 0.3794284
#>  [9,] 0.1644137 0.1644137 0.1644137
#> [10,] 3.3496794 3.3496794 3.3496794