Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.30929825 0.50954541 0.50954541
#>  [2,] 1.67286540 0.93901909 0.10112829
#>  [3,] 0.11407371 0.11407371 0.68972385
#>  [4,] 1.48621454 0.39936400 1.02350218
#>  [5,] 0.40309573 1.56228176 0.40309573
#>  [6,] 3.52133570 1.66682260 1.05881491
#>  [7,] 0.04014813 0.04014813 0.04014813
#>  [8,] 0.29965667 0.90654923 0.29965667
#>  [9,] 0.15913995 0.15913995 2.11351124
#> [10,] 0.88406316 6.38932743 0.24137542
## independence
rexmo(
    10, 3,
    c(3, 0, 0))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.01168616 0.4421304 0.16525256
#>  [2,] 0.35023095 0.5395758 0.35848881
#>  [3,] 1.03153920 1.2997909 0.86270639
#>  [4,] 1.86139365 4.0476133 0.53150007
#>  [5,] 0.45273498 0.5151104 0.02015339
#>  [6,] 3.64168109 0.4971371 1.39854921
#>  [7,] 0.69674847 0.6594870 0.22218518
#>  [8,] 0.50202712 0.7008655 1.85015400
#>  [9,] 0.80932699 1.4598485 0.99006761
#> [10,] 0.07183393 0.9225888 1.10568874
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1))
#>              [,1]        [,2]        [,3]
#>  [1,] 0.497909802 0.497909802 0.497909802
#>  [2,] 0.505947701 0.505947701 0.505947701
#>  [3,] 0.043212233 0.043212233 0.043212233
#>  [4,] 0.109368598 0.109368598 0.109368598
#>  [5,] 1.237033398 1.237033398 1.237033398
#>  [6,] 0.009915608 0.009915608 0.009915608
#>  [7,] 1.524394352 1.524394352 1.524394352
#>  [8,] 0.976894479 0.976894479 0.976894479
#>  [9,] 1.196593026 1.196593026 1.196593026
#> [10,] 0.330809779 0.330809779 0.330809779

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6196774 4.8935506 1.9003052
#>  [2,] 1.3642875 0.1809118 1.3378620
#>  [3,] 1.9449072 2.4048184 2.8455213
#>  [4,] 5.4692473 0.8909181 0.8909181
#>  [5,] 1.2861840 1.0910926 3.5702492
#>  [6,] 0.6951935 0.2958265 0.2958265
#>  [7,] 0.2179016 0.2179016 0.2179016
#>  [8,] 0.6132105 0.6132105 0.6132105
#>  [9,] 2.3206132 1.2635296 1.6226036
#> [10,] 3.2613260 1.5066498 2.8589046
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM")
#>            [,1]        [,2]       [,3]
#>  [1,] 1.9312661 0.190048515 0.79001349
#>  [2,] 0.6529949 0.368759864 1.65341916
#>  [3,] 0.9750429 0.416069565 0.92972330
#>  [4,] 0.0110427 2.836519086 0.62319074
#>  [5,] 2.5989348 1.001891623 0.88832115
#>  [6,] 0.1881753 0.009445165 0.08484926
#>  [7,] 1.1601924 1.138842534 0.28528042
#>  [8,] 0.3748848 0.022038434 2.53263123
#>  [9,] 0.5545078 0.091301569 1.92112414
#> [10,] 0.3871070 0.532452009 0.01551598
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4512098 0.4512098 0.4512098
#>  [2,] 0.1631322 0.1631322 0.1631322
#>  [3,] 0.4796647 0.4796647 0.4796647
#>  [4,] 0.9013241 0.9013241 0.9013241
#>  [5,] 1.3378400 1.3378400 1.3378400
#>  [6,] 0.1474576 0.1474576 0.1474576
#>  [7,] 3.5857563 3.5857563 3.5857563
#>  [8,] 1.8067353 1.8067353 1.8067353
#>  [9,] 0.0499683 0.0499683 0.0499683
#> [10,] 0.8213232 0.8213232 0.8213232

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3787817 0.3787817 0.37878172
#>  [2,] 2.2390500 0.1038962 0.83002816
#>  [3,] 0.2588880 4.2282044 1.76159440
#>  [4,] 1.8224281 1.8224281 1.71595704
#>  [5,] 0.4130106 1.4173380 0.41301055
#>  [6,] 1.2376299 1.6071134 0.09333631
#>  [7,] 0.2865244 0.3439790 0.34397902
#>  [8,] 0.3845580 1.4026292 1.40262916
#>  [9,] 0.8036326 1.7250743 1.72507429
#> [10,] 2.6540438 2.2612684 1.49602524
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.63477435 0.16750392 0.38064486
#>  [2,] 0.44856473 2.11116843 0.75556968
#>  [3,] 0.01428605 2.15980571 0.07974853
#>  [4,] 2.80560349 1.16137131 0.52566290
#>  [5,] 2.82252130 0.43673912 0.01717286
#>  [6,] 1.59399283 1.72606246 2.74494483
#>  [7,] 0.64532168 0.63032559 2.47235927
#>  [8,] 0.33406681 0.00424289 2.71544981
#>  [9,] 1.85152366 5.88070734 0.24174393
#> [10,] 0.56766781 0.05419533 1.77224681
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5118800 0.5118800 0.5118800
#>  [2,] 1.4234732 1.4234732 1.4234732
#>  [3,] 1.0059983 1.0059983 1.0059983
#>  [4,] 0.5035663 0.5035663 0.5035663
#>  [5,] 0.3442641 0.3442641 0.3442641
#>  [6,] 0.2689758 0.2689758 0.2689758
#>  [7,] 0.1592605 0.1592605 0.1592605
#>  [8,] 0.6474822 0.6474822 0.6474822
#>  [9,] 0.4941305 0.4941305 0.4941305
#> [10,] 0.2810564 0.2810564 0.2810564

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04444239 0.08194128 0.08194128
#>  [2,] 0.27841008 0.27841008 0.27841008
#>  [3,] 0.85598181 0.38678877 0.38678877
#>  [4,] 2.70927017 0.75563593 0.79023311
#>  [5,] 0.17560183 0.17560183 0.17560183
#>  [6,] 1.27346386 1.27346386 1.27346386
#>  [7,] 1.52447176 0.78764607 1.52447176
#>  [8,] 0.41841501 0.41841501 0.41841501
#>  [9,] 0.02483443 0.02483443 0.02483443
#> [10,] 0.93822479 0.60342065 0.93822479
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4798097 1.49625258 0.73784972
#>  [2,] 0.6521704 0.01951829 0.00906096
#>  [3,] 0.0201138 0.91128177 1.35514812
#>  [4,] 2.2141664 2.01468915 0.32622859
#>  [5,] 1.2452475 0.34942582 1.97538827
#>  [6,] 3.2361212 0.33154642 0.26936615
#>  [7,] 0.7313177 2.33603220 0.33482251
#>  [8,] 0.4366342 1.05513709 0.17104427
#>  [9,] 0.2106252 1.59064924 0.42987002
#> [10,] 0.3429674 0.83150228 0.22113149
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56149679 0.56149679 0.56149679
#>  [2,] 2.57426856 2.57426856 2.57426856
#>  [3,] 0.32465953 0.32465953 0.32465953
#>  [4,] 0.35154216 0.35154216 0.35154216
#>  [5,] 0.91713308 0.91713308 0.91713308
#>  [6,] 0.08119923 0.08119923 0.08119923
#>  [7,] 2.02624761 2.02624761 2.02624761
#>  [8,] 0.75229024 0.75229024 0.75229024
#>  [9,] 0.54250627 0.54250627 0.54250627
#> [10,] 3.12675823 3.12675823 3.12675823