Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>              [,1]       [,2]       [,3]
#>  [1,] 0.536513765 0.18672657 0.04249084
#>  [2,] 0.763110547 0.03799583 0.76311055
#>  [3,] 1.828411009 2.45326404 0.63373058
#>  [4,] 0.428159726 0.42815973 0.42815973
#>  [5,] 0.179604149 0.97644200 0.97644200
#>  [6,] 0.134035446 0.13403545 0.13403545
#>  [7,] 0.402113140 0.40211314 0.40211314
#>  [8,] 0.739310638 0.14848959 0.73931064
#>  [9,] 0.007772395 3.24055996 4.02900952
#> [10,] 0.393735618 1.05956187 0.81491896
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1546777 1.4739105 1.6350138
#>  [2,] 1.9855980 0.8123303 0.7546319
#>  [3,] 2.0549215 0.0309349 2.2594172
#>  [4,] 0.1646120 1.5407123 0.8239715
#>  [5,] 0.1158617 1.1395566 0.5603646
#>  [6,] 1.5638695 0.1512089 0.9345310
#>  [7,] 1.8648172 0.9817723 0.2413090
#>  [8,] 0.4566268 0.2117745 0.3811236
#>  [9,] 2.1318497 0.2936544 0.6007634
#> [10,] 1.8377847 2.6216879 2.5676356
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.99279049 0.99279049 0.99279049
#>  [2,] 0.77319141 0.77319141 0.77319141
#>  [3,] 2.50572901 2.50572901 2.50572901
#>  [4,] 2.24283424 2.24283424 2.24283424
#>  [5,] 0.91278608 0.91278608 0.91278608
#>  [6,] 0.40364539 0.40364539 0.40364539
#>  [7,] 1.04189610 1.04189610 1.04189610
#>  [8,] 0.08683852 0.08683852 0.08683852
#>  [9,] 0.06468575 0.06468575 0.06468575
#> [10,] 0.52065508 0.52065508 0.52065508

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 2.097465525 1.20863653 0.48185094
#>  [2,] 0.291181690 0.09827038 0.04937306
#>  [3,] 0.202997395 1.44701696 1.44701696
#>  [4,] 0.055571451 2.80431618 0.15891750
#>  [5,] 0.257707613 0.25770761 0.25770761
#>  [6,] 0.258063311 1.16440352 0.70551407
#>  [7,] 0.316432992 0.29536316 0.31643299
#>  [8,] 0.001220909 1.24726835 1.24726835
#>  [9,] 0.847382585 0.86859115 0.86859115
#> [10,] 1.326528032 0.16548551 0.09222169
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38933076 0.88947303 0.24300348
#>  [2,] 0.43740209 0.04860483 0.27772829
#>  [3,] 0.08603651 0.25914927 1.37738918
#>  [4,] 0.02819726 1.01894588 1.71472555
#>  [5,] 0.41500164 0.78892315 0.04119329
#>  [6,] 0.44292158 2.08765842 1.23130812
#>  [7,] 1.59169622 1.04419772 0.52277979
#>  [8,] 1.08130530 0.08406675 0.07011753
#>  [9,] 0.15699476 0.24821015 0.10111520
#> [10,] 2.16054861 0.81958573 2.43639068
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.95794283 0.95794283 0.95794283
#>  [2,] 1.69180305 1.69180305 1.69180305
#>  [3,] 0.21905187 0.21905187 0.21905187
#>  [4,] 0.46260039 0.46260039 0.46260039
#>  [5,] 0.08142656 0.08142656 0.08142656
#>  [6,] 0.03090017 0.03090017 0.03090017
#>  [7,] 2.46667599 2.46667599 2.46667599
#>  [8,] 1.55152497 1.55152497 1.55152497
#>  [9,] 1.68778602 1.68778602 1.68778602
#> [10,] 0.64284980 0.64284980 0.64284980

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.97923016 0.97923016 0.73512057
#>  [2,] 4.39272687 3.26998609 3.26998609
#>  [3,] 0.07605326 0.07605326 0.07605326
#>  [4,] 0.81294299 0.11902156 0.11902156
#>  [5,] 0.07270617 0.07270617 0.07270617
#>  [6,] 3.54122098 4.98529658 3.54122098
#>  [7,] 0.93324243 1.57609016 0.74728013
#>  [8,] 2.35479753 2.25482030 1.85297805
#>  [9,] 1.78686148 0.75729062 0.39740244
#> [10,] 0.08931515 0.08931515 0.08931515
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7406035 1.0267481 0.1101713
#>  [2,] 3.4159895 0.7907746 1.7427165
#>  [3,] 0.8989959 0.2676782 2.2614391
#>  [4,] 1.5863418 0.5703948 1.0689816
#>  [5,] 3.3392110 1.7540563 1.3915971
#>  [6,] 0.9325549 0.5347928 0.6822464
#>  [7,] 0.8845113 1.5649871 0.5799605
#>  [8,] 2.0946307 0.5669757 0.1988636
#>  [9,] 0.5592099 0.6224995 3.9125723
#> [10,] 1.4091664 1.1032142 0.6426058
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33608295 0.33608295 0.33608295
#>  [2,] 3.24728819 3.24728819 3.24728819
#>  [3,] 0.06722826 0.06722826 0.06722826
#>  [4,] 1.12841035 1.12841035 1.12841035
#>  [5,] 2.12287278 2.12287278 2.12287278
#>  [6,] 0.39515073 0.39515073 0.39515073
#>  [7,] 0.63715408 0.63715408 0.63715408
#>  [8,] 0.61899010 0.61899010 0.61899010
#>  [9,] 1.13258841 1.13258841 1.13258841
#> [10,] 0.78149736 0.78149736 0.78149736

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.361604230 0.361604230 0.361604230
#>  [2,] 0.300134920 0.300134920 0.300134920
#>  [3,] 0.848371941 1.883290768 0.848371941
#>  [4,] 1.447385318 0.099557987 0.099557987
#>  [5,] 2.356074513 0.798347487 2.356074513
#>  [6,] 1.996433859 2.272057042 0.867338755
#>  [7,] 0.239794858 0.239794858 0.239794858
#>  [8,] 0.152718417 1.279683688 0.250099794
#>  [9,] 0.007686953 0.007686953 0.007686953
#> [10,] 1.567940693 0.126506859 1.567940693
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>              [,1]       [,2]      [,3]
#>  [1,] 2.357231380 0.95430473 1.8688732
#>  [2,] 0.696398728 0.44242262 1.5151606
#>  [3,] 2.604282135 0.99421682 0.4570881
#>  [4,] 0.017002677 1.72653237 0.7695879
#>  [5,] 0.221524422 0.97500898 0.7462109
#>  [6,] 1.608064828 0.73475226 0.5270003
#>  [7,] 0.383903727 0.08453967 0.2023990
#>  [8,] 0.009468319 0.85514025 1.3473301
#>  [9,] 2.207639150 1.78686678 0.2709873
#> [10,] 1.472729105 0.75345327 2.9615681
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8658657 0.8658657 0.8658657
#>  [2,] 0.5316705 0.5316705 0.5316705
#>  [3,] 0.2024081 0.2024081 0.2024081
#>  [4,] 2.6292721 2.6292721 2.6292721
#>  [5,] 0.5726370 0.5726370 0.5726370
#>  [6,] 0.4596195 0.4596195 0.4596195
#>  [7,] 1.4197035 1.4197035 1.4197035
#>  [8,] 0.3086906 0.3086906 0.3086906
#>  [9,] 3.2943597 3.2943597 3.2943597
#> [10,] 1.7842775 1.7842775 1.7842775