Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]      [,2]      [,3]
#>  [1,] 1.37947204 1.7221626 0.5854707
#>  [2,] 1.89900117 1.8990012 0.6111989
#>  [3,] 1.21788252 4.5857573 2.4958189
#>  [4,] 0.12368565 0.1236856 0.1236856
#>  [5,] 2.46659659 2.1065909 1.9606630
#>  [6,] 0.26892254 0.3751305 0.2689225
#>  [7,] 0.66453467 0.7475644 0.6645347
#>  [8,] 0.05836675 2.2111041 0.6946771
#>  [9,] 1.87274843 1.8727484 1.7486803
#> [10,] 0.02045976 1.2285433 3.6444935
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3320311 1.3710056 0.32075427
#>  [2,] 1.1713783 0.0383682 1.14675701
#>  [3,] 0.5168154 2.3927067 1.09263101
#>  [4,] 1.1789852 3.4739152 1.10557230
#>  [5,] 0.1062548 0.8731818 0.27343789
#>  [6,] 3.0923781 0.5194240 0.85816629
#>  [7,] 0.2130686 0.8780629 0.06735214
#>  [8,] 0.4842390 0.4173573 0.24224278
#>  [9,] 0.2531082 0.3250745 1.14991804
#> [10,] 1.2498470 3.3672658 0.02595317
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.83216712 0.83216712 0.83216712
#>  [2,] 3.24668720 3.24668720 3.24668720
#>  [3,] 0.70518831 0.70518831 0.70518831
#>  [4,] 0.54963048 0.54963048 0.54963048
#>  [5,] 0.26229628 0.26229628 0.26229628
#>  [6,] 1.59893616 1.59893616 1.59893616
#>  [7,] 0.36284938 0.36284938 0.36284938
#>  [8,] 0.35359099 0.35359099 0.35359099
#>  [9,] 0.08320399 0.08320399 0.08320399
#> [10,] 0.32172033 0.32172033 0.32172033

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.31329691 1.31329691 1.31329691
#>  [2,] 0.61915443 1.23575602 0.50801726
#>  [3,] 2.48348737 0.68685907 1.44757436
#>  [4,] 1.50444961 0.15317660 0.09620452
#>  [5,] 0.42845427 0.47387802 0.29126884
#>  [6,] 1.81857152 0.73952543 0.19177018
#>  [7,] 0.04376543 1.01379479 0.04376543
#>  [8,] 0.49180889 0.49180889 0.49180889
#>  [9,] 0.17244554 0.17244554 0.83301276
#> [10,] 1.87152166 0.08749267 0.08749267
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.3401319 0.22287353 0.20064501
#>  [2,] 0.5258576 1.25069520 0.41752378
#>  [3,] 1.1426003 0.04050544 0.02178893
#>  [4,] 1.9114893 0.99018572 3.67200692
#>  [5,] 0.1128064 0.47897006 1.43148611
#>  [6,] 1.4384623 0.25871429 0.95906988
#>  [7,] 0.5509918 0.90272868 1.43656755
#>  [8,] 0.5378552 0.13246556 1.53126506
#>  [9,] 0.6417934 0.27011384 0.04572816
#> [10,] 1.3253953 1.83165673 0.62197460
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46900347 0.46900347 0.46900347
#>  [2,] 2.36487243 2.36487243 2.36487243
#>  [3,] 0.56155098 0.56155098 0.56155098
#>  [4,] 2.53630800 2.53630800 2.53630800
#>  [5,] 0.03698213 0.03698213 0.03698213
#>  [6,] 0.06031895 0.06031895 0.06031895
#>  [7,] 0.51016406 0.51016406 0.51016406
#>  [8,] 0.94497420 0.94497420 0.94497420
#>  [9,] 0.66036617 0.66036617 0.66036617
#> [10,] 1.17539521 1.17539521 1.17539521

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42005401 0.42005401 0.42005401
#>  [2,] 3.41641864 2.73297523 1.00073761
#>  [3,] 0.64956324 1.26685120 0.04643174
#>  [4,] 0.29987987 0.70764978 0.29987987
#>  [5,] 0.38448105 1.23305688 0.80137765
#>  [6,] 0.42522161 0.07004769 0.07004769
#>  [7,] 0.05299258 0.02079870 0.02079870
#>  [8,] 0.38840988 0.08469550 1.12167568
#>  [9,] 2.24514595 0.41547345 0.49007580
#> [10,] 0.11593000 0.11593000 0.11593000
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.37268492 0.19745608 0.85374172
#>  [2,] 0.94788013 0.17899015 0.49824083
#>  [3,] 2.44524719 0.40403781 0.29820145
#>  [4,] 0.56823838 0.17552339 0.04281005
#>  [5,] 0.16932460 1.32541231 1.22770113
#>  [6,] 1.00917753 1.23381256 0.29585313
#>  [7,] 0.73928279 0.77700601 2.27718137
#>  [8,] 0.68459701 1.24658853 0.77485763
#>  [9,] 0.02386438 0.74604514 1.33407789
#> [10,] 1.81542643 0.09418187 5.20487975
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2315156 0.2315156 0.2315156
#>  [2,] 0.9347253 0.9347253 0.9347253
#>  [3,] 0.1246727 0.1246727 0.1246727
#>  [4,] 0.9216318 0.9216318 0.9216318
#>  [5,] 2.2838720 2.2838720 2.2838720
#>  [6,] 0.1536660 0.1536660 0.1536660
#>  [7,] 0.5845356 0.5845356 0.5845356
#>  [8,] 2.0700174 2.0700174 2.0700174
#>  [9,] 1.4776064 1.4776064 1.4776064
#> [10,] 1.1263086 1.1263086 1.1263086

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.36023103 1.36023103 1.65221693
#>  [2,] 0.78553776 0.93085680 0.93085680
#>  [3,] 0.22106280 0.65225345 0.28627956
#>  [4,] 1.26428908 2.08294821 3.25011478
#>  [5,] 0.80565419 0.62611008 0.80565419
#>  [6,] 0.19804573 0.03213949 1.84204618
#>  [7,] 0.06219895 0.06219895 0.06219895
#>  [8,] 1.21516730 1.35037962 1.21516730
#>  [9,] 0.19108828 0.19108828 0.19108828
#> [10,] 0.30898583 0.30898583 0.30898583
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.41269629 1.20668420 0.5519110
#>  [2,] 0.18909503 0.52070239 1.2397749
#>  [3,] 2.47965231 0.36842142 0.5597703
#>  [4,] 1.27548162 1.12733858 2.4398507
#>  [5,] 0.06993756 0.05796901 1.2447383
#>  [6,] 0.17389576 0.12481251 0.8385308
#>  [7,] 0.80353917 1.58760428 2.3727485
#>  [8,] 0.70135178 0.21512640 0.5664314
#>  [9,] 1.53146876 3.95016715 1.9130454
#> [10,] 5.48504375 0.35060121 4.7975647
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0320872 1.0320872 1.0320872
#>  [2,] 1.0998184 1.0998184 1.0998184
#>  [3,] 3.5652300 3.5652300 3.5652300
#>  [4,] 0.2239972 0.2239972 0.2239972
#>  [5,] 1.7047234 1.7047234 1.7047234
#>  [6,] 1.9787277 1.9787277 1.9787277
#>  [7,] 0.4214590 0.4214590 0.4214590
#>  [8,] 0.6617356 0.6617356 0.6617356
#>  [9,] 0.2910710 0.2910710 0.2910710
#> [10,] 1.2846558 1.2846558 1.2846558