Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationship of exchangeable shock-size arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-counting model

The Markovian death-counting model simulates the random vector's death-counting process, which is a Markov process, until all components are dead. This defines an order statistic that is used to obtain a sample by a random permutation; see (Sloot 2022) and (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1733101 0.1733101 0.17331011
#>  [2,] 0.9368457 0.2026660 0.98051373
#>  [3,] 0.1885435 0.1885435 0.18854353
#>  [4,] 2.8550904 2.4828378 0.27470296
#>  [5,] 0.3253645 0.3253645 0.32536449
#>  [6,] 0.0294172 0.0294172 0.79498530
#>  [7,] 0.4185077 0.5914753 0.53706237
#>  [8,] 2.9185491 1.9384931 3.03985897
#>  [9,] 0.3906834 0.3906834 0.02592549
#> [10,] 0.1229192 0.6759761 0.67597607
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.7489113 0.90524881 0.28389736
#>  [2,] 0.7208594 0.74107242 2.68510167
#>  [3,] 1.1302595 0.05549046 1.80227645
#>  [4,] 0.3579772 0.10356237 0.02876517
#>  [5,] 1.2937279 1.48670305 5.73781036
#>  [6,] 0.1057949 0.60982228 2.29192023
#>  [7,] 1.0686374 0.85826729 0.83479918
#>  [8,] 0.2798746 0.78769954 0.66631343
#>  [9,] 2.1865390 0.13095003 1.48499791
#> [10,] 0.8799072 2.08176471 0.25338198
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.88633355 0.88633355 0.88633355
#>  [2,] 1.04081388 1.04081388 1.04081388
#>  [3,] 3.75810806 3.75810806 3.75810806
#>  [4,] 2.03630727 2.03630727 2.03630727
#>  [5,] 0.27788203 0.27788203 0.27788203
#>  [6,] 1.25584624 1.25584624 1.25584624
#>  [7,] 0.18355031 0.18355031 0.18355031
#>  [8,] 2.78040754 2.78040754 2.78040754
#>  [9,] 0.96700360 0.96700360 0.96700360
#> [10,] 0.04410779 0.04410779 0.04410779

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.92656098 3.08856517 0.9265610
#>  [2,] 0.80239283 0.80239283 0.8023928
#>  [3,] 0.68826964 0.58740687 0.3613996
#>  [4,] 0.85241232 3.00347963 0.1218484
#>  [5,] 0.38503448 1.66128745 2.6069629
#>  [6,] 0.06822978 0.06822978 1.9045109
#>  [7,] 3.07371496 4.95442852 4.6221446
#>  [8,] 0.40373610 0.36811146 0.4247518
#>  [9,] 0.61748613 1.30255070 1.3025507
#> [10,] 0.30007657 0.35349701 0.1178818
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]      [,2]        [,3]
#>  [1,] 0.09881098 0.2213364 3.122584689
#>  [2,] 0.09629865 0.1189554 0.524290311
#>  [3,] 3.15552997 0.9253557 1.708204957
#>  [4,] 0.72036295 1.1759266 0.268507971
#>  [5,] 0.16661659 0.3465977 0.400705212
#>  [6,] 1.68079138 1.9425320 0.900055445
#>  [7,] 1.77854509 0.2152963 0.001627893
#>  [8,] 1.14004606 1.8306195 0.181082589
#>  [9,] 1.14148157 0.3308199 0.884228919
#> [10,] 0.12067843 0.1755829 3.095592936
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1824610 0.1824610 0.1824610
#>  [2,] 2.6646207 2.6646207 2.6646207
#>  [3,] 0.4222932 0.4222932 0.4222932
#>  [4,] 1.0177464 1.0177464 1.0177464
#>  [5,] 0.6753924 0.6753924 0.6753924
#>  [6,] 0.7272906 0.7272906 0.7272906
#>  [7,] 1.4137188 1.4137188 1.4137188
#>  [8,] 2.3982228 2.3982228 2.3982228
#>  [9,] 0.9380640 0.9380640 0.9380640
#> [10,] 0.2367283 0.2367283 0.2367283

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.65159836 0.25977699 1.42984567
#>  [2,] 0.66596081 1.68578939 0.67513976
#>  [3,] 0.03308491 0.03308491 0.03308491
#>  [4,] 0.82997366 0.82997366 0.82997366
#>  [5,] 0.24116069 0.24116069 0.24116069
#>  [6,] 0.94310206 0.30189450 1.00836512
#>  [7,] 2.14462583 2.42559706 0.07240389
#>  [8,] 0.24058401 0.24058401 0.07404653
#>  [9,] 1.15753264 0.78781410 1.15753264
#> [10,] 1.92210257 0.43680136 0.46998660
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>             [,1]      [,2]        [,3]
#>  [1,] 3.75684759 0.9182379 0.349624150
#>  [2,] 4.08504474 2.2537284 0.002257809
#>  [3,] 1.66292614 4.8656922 0.725177722
#>  [4,] 0.06895228 0.6007879 1.700397431
#>  [5,] 0.28516874 0.7491690 0.270124411
#>  [6,] 0.15646704 1.2389533 0.291601104
#>  [7,] 0.91568943 0.4180694 1.048041287
#>  [8,] 0.34494374 1.0269437 0.428299634
#>  [9,] 1.17046571 0.8383817 0.527934791
#> [10,] 2.06492046 2.5808611 2.010453392
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3527401 0.3527401 0.3527401
#>  [2,] 0.9363767 0.9363767 0.9363767
#>  [3,] 2.9929656 2.9929656 2.9929656
#>  [4,] 0.2011308 0.2011308 0.2011308
#>  [5,] 0.5860842 0.5860842 0.5860842
#>  [6,] 1.0219861 1.0219861 1.0219861
#>  [7,] 0.9558936 0.9558936 0.9558936
#>  [8,] 1.0434130 1.0434130 1.0434130
#>  [9,] 0.1161993 0.1161993 0.1161993
#> [10,] 0.3952682 0.3952682 0.3952682

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 0.414971056 2.74073103 1.729908377
#>  [2,] 0.234337072 0.23433707 0.234337072
#>  [3,] 2.314711159 2.31471116 0.749497390
#>  [4,] 0.341874257 0.21983914 0.003841131
#>  [5,] 2.012194553 2.01219455 2.012194553
#>  [6,] 0.003262143 0.51765557 0.442522510
#>  [7,] 0.051440358 0.13772531 0.137725308
#>  [8,] 0.231959621 0.67031616 0.670316158
#>  [9,] 0.085031375 0.08503137 1.157356427
#> [10,] 2.145164430 1.53089180 0.204828416
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7982286 2.7783235 0.3801984
#>  [2,] 1.1985495 1.2505646 0.8562543
#>  [3,] 1.8709989 0.5364030 0.6245264
#>  [4,] 1.3313932 0.6571151 0.9620925
#>  [5,] 0.9987686 0.2471550 0.2390143
#>  [6,] 2.8880318 0.1321273 3.0930971
#>  [7,] 0.9018414 0.5188482 1.4351936
#>  [8,] 2.3211631 0.7029043 1.9129448
#>  [9,] 3.3462720 0.3543971 0.3262832
#> [10,] 1.4225504 0.6938707 1.7237543
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>               [,1]         [,2]         [,3]
#>  [1,] 0.6014604559 0.6014604559 0.6014604559
#>  [2,] 1.1766979498 1.1766979498 1.1766979498
#>  [3,] 0.0200244344 0.0200244344 0.0200244344
#>  [4,] 3.3872786088 3.3872786088 3.3872786088
#>  [5,] 1.0902417302 1.0902417302 1.0902417302
#>  [6,] 0.1099719387 0.1099719387 0.1099719387
#>  [7,] 1.7130805354 1.7130805354 1.7130805354
#>  [8,] 0.0008482421 0.0008482421 0.0008482421
#>  [9,] 3.5251631305 3.5251631305 3.5251631305
#> [10,] 0.8106830285 0.8106830285 0.8106830285