Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08171942 0.08171942 0.08171942
#>  [2,] 1.29373472 0.92746703 0.60401156
#>  [3,] 1.78039361 1.78039361 1.78039361
#>  [4,] 1.89554961 2.32764819 2.32764819
#>  [5,] 0.86775208 0.02856874 0.03245681
#>  [6,] 0.55559364 0.55559364 0.62983431
#>  [7,] 2.13149922 2.13149922 0.41592575
#>  [8,] 1.55139773 0.07382682 0.07382682
#>  [9,] 0.26693497 0.26693497 1.27457702
#> [10,] 0.38924389 1.02351433 1.02351433
rexmo(10, 3, c(3, 0, 0))          ## independence
#>              [,1]       [,2]       [,3]
#>  [1,] 1.008481557 0.38160819 1.54969801
#>  [2,] 0.939832039 1.93907903 0.65783751
#>  [3,] 0.480182389 0.27657133 0.23670129
#>  [4,] 0.001356802 0.64359624 0.83095274
#>  [5,] 2.785426921 2.40920299 0.01255004
#>  [6,] 4.070371873 0.96196386 2.38001491
#>  [7,] 0.799351319 0.10859096 2.01540239
#>  [8,] 0.776960857 0.04929243 3.69944290
#>  [9,] 1.360189229 0.45617282 0.97796612
#> [10,] 0.610805748 0.03838748 0.45995893
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2519547 0.2519547 0.2519547
#>  [2,] 0.1196111 0.1196111 0.1196111
#>  [3,] 1.2748315 1.2748315 1.2748315
#>  [4,] 0.1346091 0.1346091 0.1346091
#>  [5,] 0.2078227 0.2078227 0.2078227
#>  [6,] 0.3089128 0.3089128 0.3089128
#>  [7,] 0.2576078 0.2576078 0.2576078
#>  [8,] 1.9586784 1.9586784 1.9586784
#>  [9,] 1.2893670 1.2893670 1.2893670
#> [10,] 0.5300354 0.5300354 0.5300354

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.21551895 0.2155190 0.2155190
#>  [2,] 0.63557933 0.6355793 0.6355793
#>  [3,] 1.09762980 1.0976298 0.3989997
#>  [4,] 0.52626022 0.4276163 1.3211532
#>  [5,] 0.20394235 0.2039424 0.2039424
#>  [6,] 2.61485020 2.6148502 0.3439826
#>  [7,] 2.78172271 5.7680540 0.2178629
#>  [8,] 0.08957234 0.2985026 0.7711690
#>  [9,] 0.36144248 1.7090416 1.2710207
#> [10,] 0.58413239 1.8599128 0.5841324
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03439056 1.40991313 0.36403383
#>  [2,] 0.32704016 0.82987071 0.41069316
#>  [3,] 2.30459281 0.14621918 0.24984532
#>  [4,] 0.35328365 0.27397048 0.80485157
#>  [5,] 1.55200491 0.14758678 1.86295556
#>  [6,] 0.88646460 0.06068348 6.37228703
#>  [7,] 1.52245826 0.85054909 3.53803436
#>  [8,] 0.53536074 1.11135434 0.91813307
#>  [9,] 0.04500719 1.42102091 0.07802054
#> [10,] 0.79824015 0.64200927 0.24032036
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5284345 0.5284345 0.5284345
#>  [2,] 0.1769270 0.1769270 0.1769270
#>  [3,] 0.2291244 0.2291244 0.2291244
#>  [4,] 1.6091915 1.6091915 1.6091915
#>  [5,] 0.4020828 0.4020828 0.4020828
#>  [6,] 0.8164819 0.8164819 0.8164819
#>  [7,] 0.7987789 0.7987789 0.7987789
#>  [8,] 2.9096900 2.9096900 2.9096900
#>  [9,] 0.4780457 0.4780457 0.4780457
#> [10,] 0.5711830 0.5711830 0.5711830

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.15510512 1.60279740 1.1551051
#>  [2,] 0.52388291 0.52388291 0.5062768
#>  [3,] 1.36111136 0.21457087 1.6732733
#>  [4,] 4.69069426 0.59018000 0.5901800
#>  [5,] 0.07393040 0.07393040 0.0739304
#>  [6,] 0.58540819 0.08136824 0.2845190
#>  [7,] 0.91522630 0.91522630 1.4938586
#>  [8,] 0.04515614 0.74241873 0.7424187
#>  [9,] 0.51321445 0.51321445 0.5132145
#> [10,] 0.72628240 0.72628240 0.7262824
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.89565208 0.34105646 0.31494122
#>  [2,] 0.85199714 0.17392798 1.73942550
#>  [3,] 0.08994017 1.58076899 1.77365325
#>  [4,] 0.93180748 0.77488104 0.03768981
#>  [5,] 1.11717547 0.08920072 0.43943211
#>  [6,] 3.36129501 0.54714791 0.72618490
#>  [7,] 0.22191729 1.16342446 0.54726476
#>  [8,] 0.33712181 1.06596256 0.41356305
#>  [9,] 3.64289551 0.57592738 1.05720567
#> [10,] 0.13307710 0.44305994 1.99847825
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8536565 0.8536565 0.8536565
#>  [2,] 0.4818897 0.4818897 0.4818897
#>  [3,] 1.3508986 1.3508986 1.3508986
#>  [4,] 1.4424912 1.4424912 1.4424912
#>  [5,] 0.3038748 0.3038748 0.3038748
#>  [6,] 0.8096835 0.8096835 0.8096835
#>  [7,] 1.3527024 1.3527024 1.3527024
#>  [8,] 0.7280868 0.7280868 0.7280868
#>  [9,] 2.3725265 2.3725265 2.3725265
#> [10,] 0.9603958 0.9603958 0.9603958

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.992903628 0.29853590 2.95840623
#>  [2,] 0.008700951 1.04674633 2.40694843
#>  [3,] 0.082382964 0.08238296 0.08238296
#>  [4,] 0.252025025 0.25202502 1.31828018
#>  [5,] 2.583122645 1.81087888 1.68192696
#>  [6,] 0.148418713 0.14841871 0.04456709
#>  [7,] 0.403348667 0.40334867 0.40334867
#>  [8,] 1.362155384 0.37919967 0.35606735
#>  [9,] 0.702383138 0.70238314 0.70238314
#> [10,] 0.102182087 0.69408654 0.69408654
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]        [,2]       [,3]
#>  [1,] 1.21178264 4.089409804 0.16166409
#>  [2,] 0.91754839 2.341072979 0.86823408
#>  [3,] 0.01185954 0.342241018 0.08242056
#>  [4,] 0.16455010 2.451321693 0.94092470
#>  [5,] 1.00119114 0.009778153 1.37233650
#>  [6,] 0.34921131 2.133236378 2.97160251
#>  [7,] 2.35167866 1.819063045 1.89188267
#>  [8,] 0.31183133 1.401756471 0.04186354
#>  [9,] 0.41422146 0.064675840 1.58453176
#> [10,] 2.58786591 0.934911354 1.46793966
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.05450596 1.05450596 1.05450596
#>  [2,] 2.36369469 2.36369469 2.36369469
#>  [3,] 0.26569122 0.26569122 0.26569122
#>  [4,] 0.04826131 0.04826131 0.04826131
#>  [5,] 0.51347390 0.51347390 0.51347390
#>  [6,] 0.06847949 0.06847949 0.06847949
#>  [7,] 0.76784757 0.76784757 0.76784757
#>  [8,] 0.91116372 0.91116372 0.91116372
#>  [9,] 0.57044907 0.57044907 0.57044907
#> [10,] 1.65773375 1.65773375 1.65773375