Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]       [,2]        [,3]
#>  [1,] 2.24956965 0.39601287 0.476542254
#>  [2,] 2.53199488 2.65437238 0.339362043
#>  [3,] 2.39989266 2.39989266 2.399892660
#>  [4,] 0.36233042 2.77409318 2.639363000
#>  [5,] 0.86075448 0.90889119 0.860754484
#>  [6,] 0.03402692 0.03402692 0.004924905
#>  [7,] 0.70121412 0.81889779 0.818897786
#>  [8,] 0.98300136 0.98300136 0.805988002
#>  [9,] 0.01391968 0.01391968 0.013919682
#> [10,] 0.41513778 2.00188140 0.302562241
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6202956 0.8413291 1.7250835
#>  [2,] 0.3515848 2.6112600 0.2992968
#>  [3,] 0.3622449 0.8118616 2.2214524
#>  [4,] 1.7718425 0.3328867 1.1548697
#>  [5,] 0.3962065 0.6622826 0.8547054
#>  [6,] 0.1994369 1.5126940 3.3005701
#>  [7,] 0.5491783 1.5745857 0.1425798
#>  [8,] 4.8022607 1.9674593 1.2080865
#>  [9,] 1.5042871 2.0517026 1.7803825
#> [10,] 1.5965251 0.4920519 0.5704793
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0736223 1.0736223 1.0736223
#>  [2,] 2.2392731 2.2392731 2.2392731
#>  [3,] 1.1627572 1.1627572 1.1627572
#>  [4,] 2.3332803 2.3332803 2.3332803
#>  [5,] 0.3556859 0.3556859 0.3556859
#>  [6,] 0.2376562 0.2376562 0.2376562
#>  [7,] 1.6454777 1.6454777 1.6454777
#>  [8,] 0.8244583 0.8244583 0.8244583
#>  [9,] 0.8576396 0.8576396 0.8576396
#> [10,] 2.1785038 2.1785038 2.1785038

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39961433 0.39961433 0.39961433
#>  [2,] 1.18006784 2.57684342 2.57684342
#>  [3,] 1.07763379 0.40715781 0.55952657
#>  [4,] 0.01277388 0.01277388 0.01277388
#>  [5,] 0.08544312 0.08544312 0.08544312
#>  [6,] 0.31193578 0.31193578 0.31193578
#>  [7,] 1.70187261 0.97863487 1.08759032
#>  [8,] 0.54243985 0.54243985 0.54243985
#>  [9,] 0.56351132 1.33508474 2.22087943
#> [10,] 1.43537462 1.43537462 1.43537462
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21916883 2.14485974 2.02008664
#>  [2,] 0.12695468 0.24843169 1.84815791
#>  [3,] 2.11885012 0.06369718 0.09877829
#>  [4,] 0.07272786 5.70615589 0.41619832
#>  [5,] 1.03666237 1.18707573 0.23204662
#>  [6,] 0.22038797 3.92644703 0.15861135
#>  [7,] 0.39331591 0.01882711 0.47693893
#>  [8,] 0.78816680 1.04858845 0.71832232
#>  [9,] 1.60050140 0.50573895 1.37455722
#> [10,] 1.16906845 1.19984228 0.68967814
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15230886 0.15230886 0.15230886
#>  [2,] 1.50966695 1.50966695 1.50966695
#>  [3,] 2.09512192 2.09512192 2.09512192
#>  [4,] 2.61731973 2.61731973 2.61731973
#>  [5,] 1.41529174 1.41529174 1.41529174
#>  [6,] 0.52601072 0.52601072 0.52601072
#>  [7,] 2.70278550 2.70278550 2.70278550
#>  [8,] 0.05846408 0.05846408 0.05846408
#>  [9,] 0.08675546 0.08675546 0.08675546
#> [10,] 0.73644208 0.73644208 0.73644208

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46745687 0.46745687 0.32372401
#>  [2,] 0.02909641 0.02909641 0.02909641
#>  [3,] 0.62600861 0.55300056 0.10200085
#>  [4,] 1.32086701 1.32086701 1.32086701
#>  [5,] 1.36845480 0.64996231 2.60066712
#>  [6,] 0.70396487 0.87285563 0.21171810
#>  [7,] 0.08619564 1.14593150 0.19369352
#>  [8,] 0.18923201 1.79007694 0.18923201
#>  [9,] 0.50939778 0.63799611 2.67243762
#> [10,] 2.83041631 0.99612551 0.93830544
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2634177 0.34530405 1.11663391
#>  [2,] 0.1721767 0.10668555 0.89610085
#>  [3,] 1.8858909 0.04538887 0.96571833
#>  [4,] 0.2025964 0.70994610 0.02624747
#>  [5,] 0.3678002 0.34428274 1.10113584
#>  [6,] 3.5336390 0.98184660 0.44991980
#>  [7,] 0.2761973 0.50745106 1.52606017
#>  [8,] 1.8235959 0.36877692 2.31487952
#>  [9,] 1.8015561 0.79749176 1.87262606
#> [10,] 0.1469095 0.94630795 0.48773501
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.65700314 1.65700314 1.65700314
#>  [2,] 0.37296062 0.37296062 0.37296062
#>  [3,] 1.23588873 1.23588873 1.23588873
#>  [4,] 0.09099437 0.09099437 0.09099437
#>  [5,] 2.31228549 2.31228549 2.31228549
#>  [6,] 0.08927796 0.08927796 0.08927796
#>  [7,] 0.10502437 0.10502437 0.10502437
#>  [8,] 0.44687958 0.44687958 0.44687958
#>  [9,] 0.02338543 0.02338543 0.02338543
#> [10,] 0.45693807 0.45693807 0.45693807

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.00358520 1.00358520 1.0035852
#>  [2,] 1.38503335 1.38503335 0.8773944
#>  [3,] 0.22369619 0.22369619 0.2236962
#>  [4,] 0.48486033 0.22418713 2.6721947
#>  [5,] 2.58874340 0.06341840 0.3742726
#>  [6,] 5.70828849 0.58977052 1.4859382
#>  [7,] 1.39043966 0.03346268 0.2119076
#>  [8,] 2.38672010 3.81136678 1.6725127
#>  [9,] 0.07961047 4.05022085 1.2228028
#> [10,] 0.13119412 4.54431713 0.1311941
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4863849 0.2812067 0.3995429
#>  [2,] 0.7300578 0.4441298 1.0893960
#>  [3,] 0.4564478 0.1300588 1.7048794
#>  [4,] 1.6937388 1.1546177 1.8364814
#>  [5,] 0.1870726 2.5249104 2.4141370
#>  [6,] 0.5981893 0.1106514 0.9980711
#>  [7,] 1.9494201 1.0074970 0.2506963
#>  [8,] 1.2236766 1.4007542 1.6364532
#>  [9,] 0.7304747 1.9197222 0.1922842
#> [10,] 0.4636712 2.0031638 0.2876375
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.81512526 0.81512526 0.81512526
#>  [2,] 0.91808290 0.91808290 0.91808290
#>  [3,] 0.31824634 0.31824634 0.31824634
#>  [4,] 0.04896342 0.04896342 0.04896342
#>  [5,] 1.96262812 1.96262812 1.96262812
#>  [6,] 0.10901776 0.10901776 0.10901776
#>  [7,] 2.83182468 2.83182468 2.83182468
#>  [8,] 0.27376000 0.27376000 0.27376000
#>  [9,] 4.08831415 4.08831415 4.08831415
#> [10,] 0.01732044 0.01732044 0.01732044