Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.15885030 0.1588503 0.15885030
#>  [2,] 1.38304080 3.8147846 1.19643274
#>  [3,] 1.05790637 1.0579064 1.05790637
#>  [4,] 1.16929072 1.1692907 0.38836843
#>  [5,] 0.04067671 1.2075306 0.04067671
#>  [6,] 0.24022073 0.2402207 1.41489438
#>  [7,] 0.38367869 0.1904247 0.19042469
#>  [8,] 0.44154903 0.5494123 0.44154903
#>  [9,] 2.48151518 0.5515665 2.04789990
#> [10,] 0.26518162 0.2651816 0.26518162
rexmo(10, 3, c(3, 0, 0))          ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 1.23920514 0.1452559 0.8104151
#>  [2,] 3.10611183 2.6126798 1.4255348
#>  [3,] 0.84726178 0.2429863 0.3657959
#>  [4,] 4.49194381 1.1590889 0.4175542
#>  [5,] 1.16710072 1.5857233 1.2535854
#>  [6,] 2.25841146 0.6590139 0.6492453
#>  [7,] 0.03650204 0.1475794 0.5117247
#>  [8,] 1.11112119 0.2259113 0.8324816
#>  [9,] 1.10595076 3.0437881 0.6776535
#> [10,] 2.26636839 0.7633016 1.3453315
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9307394 0.9307394 0.9307394
#>  [2,] 0.2130737 0.2130737 0.2130737
#>  [3,] 0.1016551 0.1016551 0.1016551
#>  [4,] 2.0338554 2.0338554 2.0338554
#>  [5,] 1.2014485 1.2014485 1.2014485
#>  [6,] 0.2893647 0.2893647 0.2893647
#>  [7,] 0.3914545 0.3914545 0.3914545
#>  [8,] 0.9076400 0.9076400 0.9076400
#>  [9,] 1.1200584 1.1200584 1.1200584
#> [10,] 0.7769836 0.7769836 0.7769836

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]      [,2]        [,3]
#>  [1,] 0.09414357 0.6451129 0.006415472
#>  [2,] 3.61335425 1.7395785 0.280297730
#>  [3,] 1.34352836 1.0268696 0.499608001
#>  [4,] 0.39763645 0.1906137 0.190613677
#>  [5,] 0.73596776 1.8298922 0.735967764
#>  [6,] 0.16694168 0.1669417 0.166941677
#>  [7,] 1.73261612 0.3447067 0.327363506
#>  [8,] 0.11856735 0.1605734 0.175079694
#>  [9,] 0.05216544 0.8884291 0.052165438
#> [10,] 3.69269986 1.0361241 1.369045428
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2334179 1.2494839 0.99199021
#>  [2,] 0.4859645 0.7014472 0.01138175
#>  [3,] 0.4133786 1.0493362 1.09840106
#>  [4,] 1.2612735 2.0429437 0.80201662
#>  [5,] 0.8303112 1.9758742 0.65118703
#>  [6,] 2.2709845 1.7306517 0.91265449
#>  [7,] 0.4336813 1.6461552 0.68218752
#>  [8,] 0.2842249 0.4872301 0.02783107
#>  [9,] 0.6768089 0.2176955 0.18775159
#> [10,] 0.4542344 1.4472495 1.52819237
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41457040 0.41457040 0.41457040
#>  [2,] 0.60183187 0.60183187 0.60183187
#>  [3,] 0.82392391 0.82392391 0.82392391
#>  [4,] 0.33077944 0.33077944 0.33077944
#>  [5,] 0.98835327 0.98835327 0.98835327
#>  [6,] 0.02137117 0.02137117 0.02137117
#>  [7,] 0.02460607 0.02460607 0.02460607
#>  [8,] 0.24133665 0.24133665 0.24133665
#>  [9,] 5.42002504 5.42002504 5.42002504
#> [10,] 0.07400064 0.07400064 0.07400064

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.08527162 1.0852716 1.0852716
#>  [2,] 1.03135627 1.0313563 1.0313563
#>  [3,] 0.55233767 0.8695946 0.7783307
#>  [4,] 0.02023739 0.8829704 2.4064392
#>  [5,] 2.46380785 2.4638078 2.4638078
#>  [6,] 0.03903173 0.1137457 1.1085385
#>  [7,] 2.81689582 1.4605158 0.6605571
#>  [8,] 0.29024612 1.1001675 5.5531777
#>  [9,] 0.07389743 1.7425232 1.7425232
#> [10,] 1.97095174 0.2757001 1.5638695
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 1.51528070 0.1401060 3.80622452
#>  [2,] 1.53696584 1.2164195 1.48093671
#>  [3,] 0.80686795 0.8715300 1.62847621
#>  [4,] 0.79580484 0.1719076 0.15554784
#>  [5,] 1.62665448 0.5292522 0.03541606
#>  [6,] 0.05469317 3.8259425 0.70752651
#>  [7,] 3.15494589 0.1835887 0.45619015
#>  [8,] 0.25781659 0.1145790 2.42685107
#>  [9,] 0.35854213 0.4547602 0.32629922
#> [10,] 1.21070749 0.7761558 0.04717766
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.74938085 3.74938085 3.74938085
#>  [2,] 0.01808935 0.01808935 0.01808935
#>  [3,] 1.00849467 1.00849467 1.00849467
#>  [4,] 0.14767818 0.14767818 0.14767818
#>  [5,] 3.95506739 3.95506739 3.95506739
#>  [6,] 3.88055924 3.88055924 3.88055924
#>  [7,] 2.86019787 2.86019787 2.86019787
#>  [8,] 3.70335138 3.70335138 3.70335138
#>  [9,] 0.31473525 0.31473525 0.31473525
#> [10,] 0.12276322 0.12276322 0.12276322

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05833981 0.05833981 0.05833981
#>  [2,] 3.31876958 1.42181946 0.37128873
#>  [3,] 3.44900695 2.62886791 2.08584335
#>  [4,] 0.33727714 1.46567759 1.07879097
#>  [5,] 0.10753719 0.10753719 0.10753719
#>  [6,] 0.42424063 0.07339872 0.42424063
#>  [7,] 1.05956826 0.30762912 2.56741443
#>  [8,] 0.91847095 4.03056417 2.72665647
#>  [9,] 4.22025968 0.57334508 3.55818989
#> [10,] 0.06310355 0.06310355 0.06310355
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.70123817 0.57791135 0.06831350
#>  [2,] 0.12509269 0.33100223 0.77862454
#>  [3,] 0.15163336 0.54952484 1.87787559
#>  [4,] 0.25864914 0.73859340 4.56933202
#>  [5,] 0.38081085 0.05940075 0.05800572
#>  [6,] 0.75493385 1.49911793 1.19429001
#>  [7,] 0.01670801 0.42182425 2.62813209
#>  [8,] 2.27965236 1.71971702 1.06950142
#>  [9,] 3.61837741 0.61224214 2.89316781
#> [10,] 0.43924447 1.53814882 3.12783768
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.03023795 3.03023795 3.03023795
#>  [2,] 0.09316253 0.09316253 0.09316253
#>  [3,] 0.04620910 0.04620910 0.04620910
#>  [4,] 0.62865774 0.62865774 0.62865774
#>  [5,] 0.43231091 0.43231091 0.43231091
#>  [6,] 1.98264865 1.98264865 1.98264865
#>  [7,] 1.61526512 1.61526512 1.61526512
#>  [8,] 0.79452177 0.79452177 0.79452177
#>  [9,] 0.96211219 0.96211219 0.96211219
#> [10,] 0.94279051 0.94279051 0.94279051