Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>             [,1]      [,2]       [,3]
#>  [1,] 2.28638655 5.1969248 1.50036469
#>  [2,] 1.18286641 0.6015247 2.82172768
#>  [3,] 0.02841235 0.2455772 0.33202444
#>  [4,] 0.30787096 0.5851478 0.58514775
#>  [5,] 0.66097893 0.6609789 0.66097893
#>  [6,] 2.30871406 1.5785297 1.57852973
#>  [7,] 0.33078428 1.0232724 0.42632930
#>  [8,] 1.96215356 0.2804151 0.90755788
#>  [9,] 0.48861263 0.5131319 0.51313195
#> [10,] 2.27109600 0.4521695 0.07577466
rexmo(10, 3, c(3, 0, 0))          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.89530716 1.07984895 0.3276732
#>  [2,] 0.17159026 1.17566668 0.6262201
#>  [3,] 3.13128667 1.31243252 2.1893845
#>  [4,] 0.04169421 0.51687124 1.6320861
#>  [5,] 2.49022395 0.06995732 0.5800725
#>  [6,] 1.02213849 0.20410042 0.2341290
#>  [7,] 0.15840354 1.37295330 1.1002407
#>  [8,] 2.29169729 1.97719776 1.1354419
#>  [9,] 1.05525877 1.59107406 0.1373543
#> [10,] 1.83759480 1.17398544 0.6203822
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.9101239 2.9101239 2.9101239
#>  [2,] 0.2522818 0.2522818 0.2522818
#>  [3,] 0.3465295 0.3465295 0.3465295
#>  [4,] 1.3075168 1.3075168 1.3075168
#>  [5,] 0.1414665 0.1414665 0.1414665
#>  [6,] 0.1835876 0.1835876 0.1835876
#>  [7,] 0.0653308 0.0653308 0.0653308
#>  [8,] 3.6850239 3.6850239 3.6850239
#>  [9,] 2.9743745 2.9743745 2.9743745
#> [10,] 0.7680786 0.7680786 0.7680786

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 2.3028581 1.52571845 1.0101862
#>  [2,] 0.3051701 0.09427649 1.1642777
#>  [3,] 1.0564813 0.53987632 1.2283609
#>  [4,] 0.4342125 0.35340574 2.1612568
#>  [5,] 0.2316682 0.72322092 0.7232209
#>  [6,] 0.0652923 0.06529230 0.0652923
#>  [7,] 0.1745983 0.54122184 0.1215864
#>  [8,] 0.5596174 0.82022211 3.6804633
#>  [9,] 1.1482919 0.70788836 0.8827350
#> [10,] 1.7747529 1.73721305 1.7747529
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.28760848 0.58391808 0.0384705
#>  [2,] 0.26949920 0.17704360 0.4204575
#>  [3,] 0.81865797 0.64962423 0.8761145
#>  [4,] 0.66862918 0.04349973 0.3461309
#>  [5,] 0.16564846 0.68635164 1.9835240
#>  [6,] 2.51212330 0.57851853 0.5838518
#>  [7,] 3.44456529 0.55101822 0.8874815
#>  [8,] 0.94399235 4.65859979 0.3327713
#>  [9,] 0.08536741 0.22301304 5.8900530
#> [10,] 0.08339171 0.09822831 1.1707094
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28497708 0.28497708 0.28497708
#>  [2,] 0.33943147 0.33943147 0.33943147
#>  [3,] 0.10558217 0.10558217 0.10558217
#>  [4,] 1.00768995 1.00768995 1.00768995
#>  [5,] 0.59034297 0.59034297 0.59034297
#>  [6,] 1.69547961 1.69547961 1.69547961
#>  [7,] 0.15015016 0.15015016 0.15015016
#>  [8,] 1.51847475 1.51847475 1.51847475
#>  [9,] 0.09788872 0.09788872 0.09788872
#> [10,] 1.21642623 1.21642623 1.21642623

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.04162727 1.04276333 0.62188815
#>  [2,] 0.91493670 0.82786503 1.47753605
#>  [3,] 1.38813205 0.73333434 1.38813205
#>  [4,] 0.09276922 1.20485373 1.69623253
#>  [5,] 0.22786679 0.22786679 0.05801223
#>  [6,] 1.92875065 1.70706389 1.92875065
#>  [7,] 2.26412177 1.27513643 2.26412177
#>  [8,] 1.50862377 3.93892907 0.78340483
#>  [9,] 0.26042414 0.26042414 0.26042414
#> [10,] 0.88436106 0.02956228 1.13168777
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.60974196 0.01734151 0.5195461
#>  [2,] 3.49048745 0.24760587 2.4937788
#>  [3,] 0.39999351 0.58739683 0.1533738
#>  [4,] 1.23062224 0.34222444 3.3823807
#>  [5,] 1.28435424 1.71364052 0.2318118
#>  [6,] 0.12468083 1.70603084 0.2713671
#>  [7,] 0.04579434 2.29372863 1.4515469
#>  [8,] 0.61718012 1.01300116 0.1857487
#>  [9,] 0.10093681 0.36429859 0.9882105
#> [10,] 1.08221875 0.97619424 0.3127087
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.00210834 2.00210834 2.00210834
#>  [2,] 0.39263644 0.39263644 0.39263644
#>  [3,] 0.04130624 0.04130624 0.04130624
#>  [4,] 0.12037207 0.12037207 0.12037207
#>  [5,] 0.27767731 0.27767731 0.27767731
#>  [6,] 0.06300130 0.06300130 0.06300130
#>  [7,] 0.81854775 0.81854775 0.81854775
#>  [8,] 0.51422963 0.51422963 0.51422963
#>  [9,] 0.15600255 0.15600255 0.15600255
#> [10,] 0.26848339 0.26848339 0.26848339

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7012195 0.6251502 2.1366671
#>  [2,] 1.1730380 5.2365142 0.6677999
#>  [3,] 1.3388887 1.3388887 1.3388887
#>  [4,] 0.6834555 0.6367844 0.6834555
#>  [5,] 1.2959502 0.1639732 0.2803057
#>  [6,] 0.2139335 0.5652306 0.7308574
#>  [7,] 0.8442113 0.6970289 0.6970289
#>  [8,] 0.4269409 0.4269409 0.4269409
#>  [9,] 0.1356525 0.1356525 0.1356525
#> [10,] 1.3172174 0.2045938 0.9788467
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5656067 0.3584929 0.09038496
#>  [2,] 0.1237928 1.6621728 0.33823004
#>  [3,] 0.8507806 0.1086095 3.93414167
#>  [4,] 0.7092687 0.2188702 0.12954877
#>  [5,] 0.2785995 2.0808865 0.89676936
#>  [6,] 0.1022461 2.9455096 0.12799689
#>  [7,] 1.2016157 0.1019729 3.22008057
#>  [8,] 1.7489060 0.2513513 0.21575881
#>  [9,] 2.7279229 0.1324027 0.56301286
#> [10,] 0.3331128 0.1195916 0.55650559
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.82816440 0.82816440 0.82816440
#>  [2,] 0.26014239 0.26014239 0.26014239
#>  [3,] 2.69736733 2.69736733 2.69736733
#>  [4,] 0.44359627 0.44359627 0.44359627
#>  [5,] 0.58809744 0.58809744 0.58809744
#>  [6,] 0.11579041 0.11579041 0.11579041
#>  [7,] 0.02300282 0.02300282 0.02300282
#>  [8,] 1.01228441 1.01228441 1.01228441
#>  [9,] 1.08247486 1.08247486 1.08247486
#> [10,] 1.09752272 1.09752272 1.09752272