Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9755486 0.97554862 0.9755486
#>  [2,] 2.1609545 0.10104920 2.1609545
#>  [3,] 1.0398364 0.20517151 1.0398364
#>  [4,] 1.4756654 1.47566537 1.4756654
#>  [5,] 0.1036961 0.90191858 0.9019186
#>  [6,] 0.5393609 1.39845061 1.3984506
#>  [7,] 1.2546982 2.83125752 0.8145631
#>  [8,] 0.3094425 0.02610785 0.3094425
#>  [9,] 0.9337453 0.93374526 0.9337453
#> [10,] 0.1201370 0.12013703 0.1201370
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.50261794 2.62254400 0.8175275
#>  [2,] 0.27297630 0.65581017 3.9031524
#>  [3,] 0.77895894 0.23077486 0.6660871
#>  [4,] 1.30847166 1.09873701 0.8622911
#>  [5,] 0.04557265 1.92061475 0.4170052
#>  [6,] 1.48645925 0.06898291 0.2910510
#>  [7,] 1.71918267 0.46642898 1.0554010
#>  [8,] 0.15289023 0.58656840 1.6713833
#>  [9,] 2.60964196 0.68817260 1.0217605
#> [10,] 1.70523220 0.18737658 2.5844822
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3309110 0.3309110 0.3309110
#>  [2,] 0.9030356 0.9030356 0.9030356
#>  [3,] 1.6420832 1.6420832 1.6420832
#>  [4,] 0.3356137 0.3356137 0.3356137
#>  [5,] 1.4425578 1.4425578 1.4425578
#>  [6,] 0.4272973 0.4272973 0.4272973
#>  [7,] 1.2837923 1.2837923 1.2837923
#>  [8,] 2.1074722 2.1074722 2.1074722
#>  [9,] 0.6845670 0.6845670 0.6845670
#> [10,] 2.1866681 2.1866681 2.1866681

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2215692 0.2215692 0.2215692
#>  [2,] 0.9961470 0.9961470 0.9961470
#>  [3,] 0.7994319 0.7994319 1.1056525
#>  [4,] 1.4159201 0.0725037 0.0725037
#>  [5,] 1.5735167 0.3205975 0.3655943
#>  [6,] 1.0195905 1.0195905 2.5859827
#>  [7,] 0.5894231 0.5894231 0.5894231
#>  [8,] 1.5245905 0.9581888 1.8761645
#>  [9,] 0.8022232 2.1319410 2.1319410
#> [10,] 0.4138820 0.4138820 0.1651452
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6207729 1.5793246 0.25690967
#>  [2,] 0.1359490 0.4510098 0.43799876
#>  [3,] 1.9733529 0.2336492 0.52094374
#>  [4,] 0.7752887 0.7504896 1.48501904
#>  [5,] 0.8936912 0.5389118 1.40373904
#>  [6,] 0.6471179 1.2500525 0.05570464
#>  [7,] 3.2693417 0.1629727 0.04278916
#>  [8,] 3.2256214 0.4840151 2.93229412
#>  [9,] 0.2064106 0.1635969 0.15689315
#> [10,] 0.4117051 0.1972351 1.50441129
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.65116133 0.65116133 0.65116133
#>  [2,] 0.58296960 0.58296960 0.58296960
#>  [3,] 1.25870982 1.25870982 1.25870982
#>  [4,] 2.59243426 2.59243426 2.59243426
#>  [5,] 1.07750672 1.07750672 1.07750672
#>  [6,] 0.88411333 0.88411333 0.88411333
#>  [7,] 0.85611773 0.85611773 0.85611773
#>  [8,] 0.68123179 0.68123179 0.68123179
#>  [9,] 0.44829814 0.44829814 0.44829814
#> [10,] 0.04642791 0.04642791 0.04642791

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.43084355 0.4308436 0.43084355
#>  [2,] 0.88899017 1.6585397 1.59959983
#>  [3,] 0.30212097 0.3021210 0.30212097
#>  [4,] 0.58835666 0.5883567 0.31587058
#>  [5,] 0.92370555 1.1184901 0.01389528
#>  [6,] 0.72876350 0.7287635 0.72876350
#>  [7,] 0.06325762 0.5619430 1.71487344
#>  [8,] 0.15340593 2.6182335 0.74083075
#>  [9,] 0.74296445 0.7429645 0.74296445
#> [10,] 3.03571379 2.1434155 3.19997040
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.75974740 0.45763687 0.77597639
#>  [2,] 2.30702775 0.02175808 0.21264012
#>  [3,] 0.57121760 1.83775396 1.66506105
#>  [4,] 0.74818410 0.73727990 0.10011306
#>  [5,] 2.29365157 1.82652080 0.08536791
#>  [6,] 1.80677557 0.09212463 0.45871085
#>  [7,] 0.01965100 0.75284519 0.07828083
#>  [8,] 1.81618464 0.97460970 0.93436859
#>  [9,] 0.02513142 2.15077852 1.02664523
#> [10,] 0.56294798 1.45727153 0.48815906
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.564565248 0.564565248 0.564565248
#>  [2,] 0.846988149 0.846988149 0.846988149
#>  [3,] 0.126699006 0.126699006 0.126699006
#>  [4,] 0.311447337 0.311447337 0.311447337
#>  [5,] 0.604450407 0.604450407 0.604450407
#>  [6,] 0.492215966 0.492215966 0.492215966
#>  [7,] 1.341293426 1.341293426 1.341293426
#>  [8,] 0.897392473 0.897392473 0.897392473
#>  [9,] 0.748010481 0.748010481 0.748010481
#> [10,] 0.002419402 0.002419402 0.002419402

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.8676363 0.2753399 1.86763626
#>  [2,] 0.5749009 0.5749009 0.57490091
#>  [3,] 1.5510601 3.1836705 0.01938808
#>  [4,] 0.9983347 1.4264160 0.59194114
#>  [5,] 0.9637964 0.5475221 0.96379636
#>  [6,] 0.9778561 1.4358806 0.97785608
#>  [7,] 0.7654990 0.7654990 0.15235079
#>  [8,] 0.2674623 0.2674623 0.26746231
#>  [9,] 0.3196473 0.3196473 0.31964726
#> [10,] 0.9900306 0.5731438 0.57314384
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 2.8263302 0.5773776 1.02680783
#>  [2,] 0.4562137 1.5424222 0.32446519
#>  [3,] 0.5327006 0.2672791 1.10326051
#>  [4,] 4.1314494 0.4433925 1.38547261
#>  [5,] 1.7492959 1.2427920 3.50253921
#>  [6,] 0.2325141 0.9129006 0.34021800
#>  [7,] 0.5845521 1.5958843 0.39315993
#>  [8,] 2.5195828 1.1938785 1.93539402
#>  [9,] 1.2530795 1.1426925 1.31325299
#> [10,] 2.1862455 4.2576693 0.01596324
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.36137635 2.36137635 2.36137635
#>  [2,] 1.90133530 1.90133530 1.90133530
#>  [3,] 0.02902261 0.02902261 0.02902261
#>  [4,] 0.74432215 0.74432215 0.74432215
#>  [5,] 0.74317853 0.74317853 0.74317853
#>  [6,] 2.38877300 2.38877300 2.38877300
#>  [7,] 1.06702723 1.06702723 1.06702723
#>  [8,] 1.00206453 1.00206453 1.00206453
#>  [9,] 2.30844777 2.30844777 2.30844777
#> [10,] 0.99912492 0.99912492 0.99912492