Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.03910310 2.7629531 0.75005638
#>  [2,] 1.49167079 0.2813662 0.28136623
#>  [3,] 1.31018563 0.6964609 0.43799461
#>  [4,] 0.73271421 2.3488835 0.53710024
#>  [5,] 0.37293535 0.3729353 1.18603495
#>  [6,] 0.05613014 1.0569323 1.05693228
#>  [7,] 1.45981871 0.7180965 1.94918599
#>  [8,] 0.25116916 0.1084164 0.05423552
#>  [9,] 2.04163377 1.8835688 0.06859072
#> [10,] 0.75895378 1.8640239 0.22487811
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8008922 0.66170690 4.3722239
#>  [2,] 1.6343091 0.38077200 1.0882514
#>  [3,] 2.3565694 1.28385620 0.5455765
#>  [4,] 4.4324292 0.26236732 0.8678752
#>  [5,] 1.4654421 0.02204145 0.1251700
#>  [6,] 0.3003466 1.30426794 0.6383970
#>  [7,] 1.6839657 0.51431552 1.5345917
#>  [8,] 1.0092190 1.04751368 0.4096108
#>  [9,] 1.1715560 1.45752664 1.0205629
#> [10,] 2.0386997 0.72930421 2.8678189
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8010010 1.8010010 1.8010010
#>  [2,] 1.1420193 1.1420193 1.1420193
#>  [3,] 0.7962134 0.7962134 0.7962134
#>  [4,] 1.6543221 1.6543221 1.6543221
#>  [5,] 0.7666708 0.7666708 0.7666708
#>  [6,] 0.8172957 0.8172957 0.8172957
#>  [7,] 1.7995371 1.7995371 1.7995371
#>  [8,] 1.4092542 1.4092542 1.4092542
#>  [9,] 0.2051142 0.2051142 0.2051142
#> [10,] 1.3056977 1.3056977 1.3056977

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3327893 1.8463960 2.03739999
#>  [2,] 2.0387610 2.6941905 1.60751608
#>  [3,] 4.4728975 4.3247456 6.36618576
#>  [4,] 1.4457256 0.4035308 1.70644443
#>  [5,] 0.3935227 0.7327385 0.89011398
#>  [6,] 0.7587094 0.7587094 0.75870941
#>  [7,] 0.2282501 0.8864157 1.08976315
#>  [8,] 1.1374494 0.5458898 0.44105995
#>  [9,] 4.7789319 2.6298921 0.02503268
#> [10,] 0.2453469 0.2453469 0.24534688
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.09791133 0.15343944 0.03049748
#>  [2,] 0.09949029 0.57353214 2.38079879
#>  [3,] 0.01172541 0.08587196 1.89844401
#>  [4,] 0.96635425 0.18760874 0.43994173
#>  [5,] 1.27961420 0.53176667 1.09265209
#>  [6,] 3.43076266 0.04852240 0.48006303
#>  [7,] 0.31090320 0.13701916 0.14486861
#>  [8,] 2.68751548 0.03317181 0.15266962
#>  [9,] 1.55928538 0.72775721 0.42948540
#> [10,] 0.37594608 0.34852078 0.05872754
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.20510555 1.20510555 1.20510555
#>  [2,] 1.02483234 1.02483234 1.02483234
#>  [3,] 1.77212364 1.77212364 1.77212364
#>  [4,] 1.09698349 1.09698349 1.09698349
#>  [5,] 3.80357822 3.80357822 3.80357822
#>  [6,] 0.05365136 0.05365136 0.05365136
#>  [7,] 0.16769803 0.16769803 0.16769803
#>  [8,] 0.48135053 0.48135053 0.48135053
#>  [9,] 0.70475136 0.70475136 0.70475136
#> [10,] 1.03597174 1.03597174 1.03597174

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.00381420 0.36592866 0.97162866
#>  [2,] 1.06985284 0.11907340 1.40742657
#>  [3,] 2.94869755 0.03276066 2.94869755
#>  [4,] 2.49544406 0.19094059 2.38622262
#>  [5,] 0.21286164 0.21286164 0.21286164
#>  [6,] 0.69315327 0.69315327 0.69315327
#>  [7,] 3.81024245 0.41318315 0.41318315
#>  [8,] 0.04108176 0.04108176 0.04108176
#>  [9,] 0.58382505 0.58382505 0.58382505
#> [10,] 0.45075344 5.30142055 1.98569401
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 1.46214062 0.008105964 0.08081834
#>  [2,] 3.98267476 0.183257910 1.65794280
#>  [3,] 0.45538391 1.354754424 0.20113519
#>  [4,] 0.08761483 2.391346662 0.03409392
#>  [5,] 0.59531697 3.964295250 2.58230709
#>  [6,] 1.59794195 0.098203301 1.64987751
#>  [7,] 1.68341424 1.843050741 0.12115358
#>  [8,] 1.98666311 0.622432039 1.65799028
#>  [9,] 0.09291632 0.614727829 0.41817654
#> [10,] 1.03029097 0.244143756 1.38121421
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 3.2301963 3.2301963 3.2301963
#>  [2,] 1.7779116 1.7779116 1.7779116
#>  [3,] 0.5877495 0.5877495 0.5877495
#>  [4,] 0.1600403 0.1600403 0.1600403
#>  [5,] 1.2080304 1.2080304 1.2080304
#>  [6,] 3.5541174 3.5541174 3.5541174
#>  [7,] 1.6128778 1.6128778 1.6128778
#>  [8,] 0.0718742 0.0718742 0.0718742
#>  [9,] 0.7908461 0.7908461 0.7908461
#> [10,] 0.7902583 0.7902583 0.7902583

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5662509 0.3408078 0.5390147
#>  [2,] 0.7344016 0.5140222 0.7344016
#>  [3,] 0.1887906 0.4211000 0.4211000
#>  [4,] 4.7817334 1.7634427 0.1408955
#>  [5,] 0.6025414 1.7129319 0.2857823
#>  [6,] 0.8852973 0.8852973 0.8358194
#>  [7,] 0.5367237 0.4862974 0.2656157
#>  [8,] 0.7378198 1.4431252 1.7019196
#>  [9,] 1.4835564 1.3594465 0.2988343
#> [10,] 0.9701228 1.4885276 0.9701228
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7730945 0.05547609 0.2703245
#>  [2,] 1.1060958 0.63150267 2.2907373
#>  [3,] 2.5665189 0.50173918 0.9632740
#>  [4,] 0.7823600 1.12694827 2.2649882
#>  [5,] 0.8067667 3.28727608 1.2445191
#>  [6,] 0.8971160 0.06153586 0.2102764
#>  [7,] 1.2711529 1.40200703 1.0159973
#>  [8,] 1.9787080 0.95072984 0.8044869
#>  [9,] 0.3622451 0.35666643 0.8077635
#> [10,] 0.5694748 0.15032386 0.2333111
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.86521892 0.86521892 0.86521892
#>  [2,] 1.07462874 1.07462874 1.07462874
#>  [3,] 2.56294570 2.56294570 2.56294570
#>  [4,] 2.14600700 2.14600700 2.14600700
#>  [5,] 0.18165171 0.18165171 0.18165171
#>  [6,] 0.65716015 0.65716015 0.65716015
#>  [7,] 1.96309441 1.96309441 1.96309441
#>  [8,] 0.05160971 0.05160971 0.05160971
#>  [9,] 0.07495551 0.07495551 0.07495551
#> [10,] 0.59172160 0.59172160 0.59172160