Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationship of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9892938 2.70706228 1.4965479
#>  [2,] 0.1255569 1.06187231 0.9141439
#>  [3,] 1.1201035 1.12010351 0.5994747
#>  [4,] 1.1873050 1.18730497 1.5536284
#>  [5,] 0.2384664 0.89432186 0.6586136
#>  [6,] 2.0875379 0.81164473 0.8116447
#>  [7,] 0.6270901 0.02059166 4.9730146
#>  [8,] 3.0385398 3.10744402 3.1074440
#>  [9,] 0.6353838 0.63538378 0.6353838
#> [10,] 2.2476264 2.18295592 0.7572122
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.59840110 0.1049803 0.13066758
#>  [2,] 0.25385505 0.4603088 0.69402862
#>  [3,] 0.96295470 0.3965292 0.86019072
#>  [4,] 1.87275477 0.3013698 0.23823128
#>  [5,] 0.15282829 0.6699738 0.56488712
#>  [6,] 0.17991566 1.0667001 0.21769126
#>  [7,] 0.10483393 3.6531172 0.01960146
#>  [8,] 0.37404751 1.6262110 1.58557837
#>  [9,] 0.07214976 2.1540489 0.49493593
#> [10,] 0.14959789 1.3120407 3.52196264
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08954005 0.08954005 0.08954005
#>  [2,] 0.82101753 0.82101753 0.82101753
#>  [3,] 2.51723319 2.51723319 2.51723319
#>  [4,] 0.74262737 0.74262737 0.74262737
#>  [5,] 1.72599862 1.72599862 1.72599862
#>  [6,] 0.07547171 0.07547171 0.07547171
#>  [7,] 0.88345994 0.88345994 0.88345994
#>  [8,] 0.38966611 0.38966611 0.38966611
#>  [9,] 2.84744221 2.84744221 2.84744221
#> [10,] 2.28731970 2.28731970 2.28731970

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.44912074 1.54948512 3.03662227
#>  [2,] 3.41761496 0.45869247 0.45869247
#>  [3,] 0.48984025 0.48984025 0.48984025
#>  [4,] 0.05748718 0.05748718 0.05748718
#>  [5,] 0.14473048 0.14473048 0.14473048
#>  [6,] 0.17066190 1.67603924 0.36040457
#>  [7,] 1.43296221 1.02663028 0.96265848
#>  [8,] 0.65949032 0.93771816 0.22087383
#>  [9,] 0.28699016 0.01771606 0.28699016
#> [10,] 0.12645487 0.12645487 0.12645487
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.49750073 0.2975480 2.94208477
#>  [2,] 1.70779525 0.3361682 1.09469732
#>  [3,] 0.61863665 0.9051763 0.27246744
#>  [4,] 0.32771079 0.8848463 0.01612511
#>  [5,] 1.11104775 1.1332225 0.21380761
#>  [6,] 0.08941564 0.7831808 0.39873506
#>  [7,] 0.98149353 0.9030860 1.17453532
#>  [8,] 0.74481994 1.2514963 2.34418184
#>  [9,] 0.11027859 1.0981273 1.34175332
#> [10,] 2.02684338 0.1029716 0.56986389
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.80816398 1.80816398 1.80816398
#>  [2,] 0.03066247 0.03066247 0.03066247
#>  [3,] 0.59480466 0.59480466 0.59480466
#>  [4,] 2.42257507 2.42257507 2.42257507
#>  [5,] 0.02627880 0.02627880 0.02627880
#>  [6,] 1.49949175 1.49949175 1.49949175
#>  [7,] 1.01314487 1.01314487 1.01314487
#>  [8,] 1.35325843 1.35325843 1.35325843
#>  [9,] 0.99163227 0.99163227 0.99163227
#> [10,] 0.42974212 0.42974212 0.42974212

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.35032370 0.49735754 2.8258553
#>  [2,] 1.44966309 0.92086326 0.2275155
#>  [3,] 0.06588980 0.06588980 1.5621840
#>  [4,] 0.66209291 2.36294767 0.5093792
#>  [5,] 3.41693055 2.71485676 2.7996126
#>  [6,] 1.18160195 0.41420618 0.9014830
#>  [7,] 0.58902807 0.01629522 1.2011700
#>  [8,] 3.70443249 2.07979588 3.7044325
#>  [9,] 0.08048095 0.08048095 1.0678606
#> [10,] 0.44395125 0.44395125 0.4439512
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.71871978 3.97585853 2.81199904
#>  [2,] 1.55940391 4.98276825 1.98376408
#>  [3,] 0.85757162 0.53012518 0.51381348
#>  [4,] 0.13459842 0.88453574 1.11709336
#>  [5,] 0.31294720 1.03239547 2.61969486
#>  [6,] 0.34301301 0.06951085 1.07096187
#>  [7,] 0.01856991 0.33209666 1.54972245
#>  [8,] 0.24310033 0.71084242 1.14629572
#>  [9,] 0.61032922 0.85360948 0.99873607
#> [10,] 0.29137741 1.60263638 0.01006345
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2156803 0.2156803 0.2156803
#>  [2,] 1.5863757 1.5863757 1.5863757
#>  [3,] 0.3333308 0.3333308 0.3333308
#>  [4,] 0.7659731 0.7659731 0.7659731
#>  [5,] 0.8204958 0.8204958 0.8204958
#>  [6,] 1.8723279 1.8723279 1.8723279
#>  [7,] 0.7511116 0.7511116 0.7511116
#>  [8,] 0.6574324 0.6574324 0.6574324
#>  [9,] 0.3999442 0.3999442 0.3999442
#> [10,] 2.2338032 2.2338032 2.2338032

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.47456760 0.4745676 0.4745676
#>  [2,] 2.63539570 3.0071297 2.2899140
#>  [3,] 2.74443234 0.1317810 0.1317810
#>  [4,] 0.78451914 0.7845191 0.0581099
#>  [5,] 0.74371327 0.7437133 0.7437133
#>  [6,] 0.96492749 0.9649275 1.3913752
#>  [7,] 0.04070342 1.0345662 1.0345662
#>  [8,] 0.25275507 5.3967763 2.9400742
#>  [9,] 1.10276327 0.1375234 2.4923837
#> [10,] 1.64686663 1.6468666 1.6468666
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.21540453 0.5499197 0.06844118
#>  [2,] 2.00818747 0.2477633 2.69145336
#>  [3,] 0.38140147 0.1250359 0.70693911
#>  [4,] 0.78971625 0.4981457 0.28785869
#>  [5,] 0.40737287 1.1748172 0.51487974
#>  [6,] 0.36327827 2.4242514 0.21932664
#>  [7,] 0.09633111 0.2182969 0.15583952
#>  [8,] 1.58169814 0.9859893 0.11174040
#>  [9,] 2.58442493 0.2687974 0.56059644
#> [10,] 0.22945617 0.0494611 3.05150287
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.43834898 3.43834898 3.43834898
#>  [2,] 0.02388589 0.02388589 0.02388589
#>  [3,] 6.06201564 6.06201564 6.06201564
#>  [4,] 0.78747526 0.78747526 0.78747526
#>  [5,] 0.13510453 0.13510453 0.13510453
#>  [6,] 0.83763627 0.83763627 0.83763627
#>  [7,] 3.24769309 3.24769309 3.24769309
#>  [8,] 1.85244550 1.85244550 1.85244550
#>  [9,] 0.19459137 0.19459137 0.19459137
#> [10,] 1.51983611 1.51983611 1.51983611