Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3040951 0.1689512 0.3040951
#>  [2,] 1.0597560 1.5168298 1.8509399
#>  [3,] 0.3166725 1.4311630 0.3166725
#>  [4,] 0.6263029 0.8267506 0.4016470
#>  [5,] 0.1744947 0.1744947 0.1480003
#>  [6,] 3.8720777 2.1400878 1.6882639
#>  [7,] 2.5818337 1.0213909 1.0727545
#>  [8,] 1.2449259 1.2449259 1.2449259
#>  [9,] 1.0762261 2.8755744 0.4185312
#> [10,] 0.2084681 0.2084681 0.2084681
rexmo(10, 3, c(3, 0, 0))          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43508635 0.80501885 0.79081332
#>  [2,] 0.08377795 0.36885822 0.12913176
#>  [3,] 2.17136533 0.04342258 0.25543343
#>  [4,] 0.54556259 0.60193701 1.29687463
#>  [5,] 1.05042894 0.35570718 1.45782198
#>  [6,] 0.85146155 1.01970577 1.21614078
#>  [7,] 0.32844034 0.99085845 0.62803371
#>  [8,] 0.40458832 0.47508661 0.77361976
#>  [9,] 0.48755463 0.66567899 2.57258576
#> [10,] 0.19063034 1.05919744 0.04291181
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.88898129 1.88898129 1.88898129
#>  [2,] 0.47026509 0.47026509 0.47026509
#>  [3,] 0.18100340 0.18100340 0.18100340
#>  [4,] 0.15202840 0.15202840 0.15202840
#>  [5,] 2.70306748 2.70306748 2.70306748
#>  [6,] 1.52478465 1.52478465 1.52478465
#>  [7,] 0.02574678 0.02574678 0.02574678
#>  [8,] 1.53999246 1.53999246 1.53999246
#>  [9,] 0.00412451 0.00412451 0.00412451
#> [10,] 0.60267125 0.60267125 0.60267125

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 5.47649728 3.20738497 3.5413747
#>  [2,] 0.26816940 0.26816940 0.2681694
#>  [3,] 0.31817808 0.31817808 0.3181781
#>  [4,] 1.39322002 1.51633488 0.4016386
#>  [5,] 0.80892245 1.00158824 1.0015882
#>  [6,] 1.05531724 0.10063366 1.0124178
#>  [7,] 0.04614953 4.03094316 1.5148984
#>  [8,] 2.65091802 0.09537461 7.5648305
#>  [9,] 0.43292107 0.43292107 0.8230840
#> [10,] 2.18983551 3.40065038 1.6858443
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.64509954 1.7478321 1.3782749
#>  [2,] 0.30838069 0.6403060 0.2451046
#>  [3,] 0.04211762 1.9135442 0.8473018
#>  [4,] 1.49641964 0.1610214 0.1374510
#>  [5,] 1.26030968 0.7766153 0.1368264
#>  [6,] 0.94046732 2.2760702 1.3741349
#>  [7,] 0.30782508 0.1844143 1.7894420
#>  [8,] 0.48547401 0.4970336 0.4088441
#>  [9,] 1.87531608 1.8435359 0.2588724
#> [10,] 0.03673357 0.2604531 0.3135096
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2424547 0.2424547 0.2424547
#>  [2,] 0.7448031 0.7448031 0.7448031
#>  [3,] 2.2483164 2.2483164 2.2483164
#>  [4,] 0.9295371 0.9295371 0.9295371
#>  [5,] 1.0085785 1.0085785 1.0085785
#>  [6,] 1.7012777 1.7012777 1.7012777
#>  [7,] 0.5966434 0.5966434 0.5966434
#>  [8,] 1.9264673 1.9264673 1.9264673
#>  [9,] 1.4349277 1.4349277 1.4349277
#> [10,] 1.1482203 1.1482203 1.1482203

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.051673012 0.05167301 0.05167301
#>  [2,] 0.226555253 0.60337276 0.33980945
#>  [3,] 0.041446445 1.61797115 0.04144644
#>  [4,] 1.267706777 1.18783984 0.55252253
#>  [5,] 0.325039579 0.40665007 0.32503958
#>  [6,] 0.440757398 0.44075740 0.44075740
#>  [7,] 1.324616852 1.15125696 0.44584000
#>  [8,] 2.216388436 2.21638844 2.21638844
#>  [9,] 0.007635925 0.48410740 0.42011686
#> [10,] 0.298525933 0.51454404 0.51454404
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.95334239 1.4075892 0.7079832
#>  [2,] 1.02974260 0.2853046 1.4952779
#>  [3,] 2.59710046 0.1011125 0.4572534
#>  [4,] 0.05083939 1.6729628 0.7457807
#>  [5,] 0.65130448 3.3165372 0.1690687
#>  [6,] 0.66577531 1.3284876 0.5738329
#>  [7,] 1.43114157 0.5251145 0.1066416
#>  [8,] 0.43388989 0.7520198 0.8948818
#>  [9,] 0.56576585 0.5457167 2.2076358
#> [10,] 2.36379960 0.2205456 0.6400546
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08178056 0.08178056 0.08178056
#>  [2,] 1.81408659 1.81408659 1.81408659
#>  [3,] 0.67123758 0.67123758 0.67123758
#>  [4,] 0.54483051 0.54483051 0.54483051
#>  [5,] 0.21233085 0.21233085 0.21233085
#>  [6,] 2.93219294 2.93219294 2.93219294
#>  [7,] 0.44914015 0.44914015 0.44914015
#>  [8,] 0.18165743 0.18165743 0.18165743
#>  [9,] 1.37526752 1.37526752 1.37526752
#> [10,] 1.31140288 1.31140288 1.31140288

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.1580340 0.12051912 0.52351986
#>  [2,] 1.7627580 0.53309767 2.53996080
#>  [3,] 1.0877770 0.09084864 0.09084864
#>  [4,] 1.3447621 1.34476210 0.25639989
#>  [5,] 0.2432415 0.24324149 0.24324149
#>  [6,] 1.2531247 1.71603615 1.92823521
#>  [7,] 1.0293966 1.02939656 0.87921575
#>  [8,] 0.4432880 0.44328802 0.44328802
#>  [9,] 0.7270246 1.46087576 0.23855765
#> [10,] 0.8994224 0.89942236 2.71059478
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.71478470 0.7304161 2.68535917
#>  [2,] 1.04230501 1.2980405 1.28335850
#>  [3,] 0.63808045 2.5463582 0.66928234
#>  [4,] 0.47408588 0.9167633 2.03871640
#>  [5,] 1.21393646 1.8036602 0.26615156
#>  [6,] 0.36268138 2.6139520 1.58083612
#>  [7,] 0.29868290 0.7635701 1.50935166
#>  [8,] 0.07440108 3.2487958 0.19886561
#>  [9,] 0.12235777 0.3623675 1.28315822
#> [10,] 0.75197927 0.6184603 0.03318307
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40497812 0.40497812 0.40497812
#>  [2,] 0.04801816 0.04801816 0.04801816
#>  [3,] 0.73657752 0.73657752 0.73657752
#>  [4,] 0.08579793 0.08579793 0.08579793
#>  [5,] 4.34723508 4.34723508 4.34723508
#>  [6,] 0.70925800 0.70925800 0.70925800
#>  [7,] 1.04224768 1.04224768 1.04224768
#>  [8,] 0.55449475 0.55449475 0.55449475
#>  [9,] 0.32453003 0.32453003 0.32453003
#> [10,] 2.25218148 2.25218148 2.25218148