Sample from an exchangeable Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size arrival rates.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size arrival intensitiesex_intensities.

Parametrization

The exchangeable Marshall–Olkin distribution has the following property: a shock arrival intensity \(\lambda_I\) for a set \(I\) only depends on the cardinality of \(I\), i.e., the exchangeable shock arrival intensities are defined for \(I\) with \(i = \lvert I\rvert\) by \(\lambda_i = \lambda_I\), see (see Sec. 3.2 Mai and Scherer 2017) . The exchangeable Marshall–Olkin distribution is parametrized by the exchangeable shock-size arrival intensities defined by $$ \eta_i = \binom{d}{i} \lambda_i , \quad i \in {\{ 1, \ldots, d \}} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The ESM and AM simulation algorithms for the general Marshall–Olkin distribution can be used; for this, the exchangeable shock-size arrival intensities are converted to the corresponding shock arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rexmo(10, 3, c(1.2, 0.3, 0.4))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2622758 0.2622758 0.2410682
#>  [2,] 0.8820979 0.8820979 0.7919674
#>  [3,] 1.5200031 1.8959821 2.4052115
#>  [4,] 5.6596258 2.9761266 4.8781700
#>  [5,] 0.1591860 2.3553982 0.9750104
#>  [6,] 0.4198689 0.4198689 0.4198689
#>  [7,] 1.9801351 1.4648855 1.5736583
#>  [8,] 0.4675945 0.4675945 0.4675945
#>  [9,] 0.9525670 1.0908686 0.7247127
#> [10,] 1.3857809 2.0154148 0.1100043
rexmo(10, 3, c(3, 0, 0))          ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5789284 0.1165352 0.1737120
#>  [2,] 0.6273972 0.5917747 1.5764016
#>  [3,] 1.5370047 1.4893611 1.7751226
#>  [4,] 1.2396511 1.5966276 0.1840400
#>  [5,] 0.8033410 0.8079853 0.3088655
#>  [6,] 0.7465569 0.6018163 0.3004492
#>  [7,] 0.9925346 0.6084240 3.6774140
#>  [8,] 2.0147946 2.3836279 2.9874238
#>  [9,] 2.1728481 2.5371109 0.3958424
#> [10,] 1.3258751 1.0852391 0.3157392
rexmo(10, 3, c(0, 0, 1))          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37557326 0.37557326 0.37557326
#>  [2,] 0.13611977 0.13611977 0.13611977
#>  [3,] 0.39732079 0.39732079 0.39732079
#>  [4,] 1.26811703 1.26811703 1.26811703
#>  [5,] 1.43996097 1.43996097 1.43996097
#>  [6,] 2.76436484 2.76436484 2.76436484
#>  [7,] 0.08347254 0.08347254 0.08347254
#>  [8,] 0.91695712 0.91695712 0.91695712
#>  [9,] 1.08697439 1.08697439 1.08697439
#> [10,] 3.84003014 3.84003014 3.84003014

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,]  0.53424891 0.07866340 1.03069569
#>  [2,]  0.05454759 0.14517220 1.47601239
#>  [3,]  0.06608259 0.06608259 0.06608259
#>  [4,]  0.81841367 0.72070682 0.81841367
#>  [5,] 11.60356118 2.85414294 0.31346639
#>  [6,]  1.95553386 0.17074271 0.17074271
#>  [7,]  1.40363262 0.04703766 0.04703766
#>  [8,]  1.13006257 1.13006257 0.61580587
#>  [9,]  0.50916664 0.28028925 0.50916664
#> [10,]  1.60065176 1.60065176 1.72120757
rexmo(10, 3, c(3, 0, 0), method = "MDCM")          ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.63844066 1.04300474 0.3154808
#>  [2,] 0.13995102 0.40027034 0.2565090
#>  [3,] 0.11409861 0.51169699 0.2769425
#>  [4,] 0.22195089 0.20828320 0.5302511
#>  [5,] 0.02502959 0.02137802 0.7698678
#>  [6,] 0.57008766 1.23510625 0.4746647
#>  [7,] 2.06180446 1.04798927 0.5273681
#>  [8,] 0.72146133 0.69461230 0.3653632
#>  [9,] 0.91385879 0.92599591 0.7268215
#> [10,] 0.80832741 0.01071109 1.1608281
rexmo(10, 3, c(0, 0, 1), method = "MDCM")          ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9923269 0.9923269 0.9923269
#>  [2,] 0.9856149 0.9856149 0.9856149
#>  [3,] 2.0455814 2.0455814 2.0455814
#>  [4,] 0.7126431 0.7126431 0.7126431
#>  [5,] 0.3426729 0.3426729 0.3426729
#>  [6,] 0.9598300 0.9598300 0.9598300
#>  [7,] 2.5230597 2.5230597 2.5230597
#>  [8,] 0.1399707 0.1399707 0.1399707
#>  [9,] 0.3956437 0.3956437 0.3956437
#> [10,] 0.1293918 0.1293918 0.1293918

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.48364089 0.48364089 0.4836409
#>  [2,] 0.03601102 0.03601102 1.7184357
#>  [3,] 0.47508440 0.61745807 0.4659214
#>  [4,] 3.67582007 1.46536814 1.4653681
#>  [5,] 0.72417663 2.11601934 2.3960321
#>  [6,] 0.22127412 0.02266004 1.5932046
#>  [7,] 0.64245226 0.64245226 0.6424523
#>  [8,] 0.15804718 0.15804718 0.0521973
#>  [9,] 0.30540821 0.81346190 2.3695271
#> [10,] 0.88433170 0.88433170 0.8843317
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
#>            [,1]       [,2]         [,3]
#>  [1,] 0.2546447 0.02088411 0.0594654014
#>  [2,] 0.1719427 2.04355464 0.7294814752
#>  [3,] 1.7585819 1.50682317 0.0003410636
#>  [4,] 0.2303619 2.99243930 2.2514223773
#>  [5,] 0.1556194 0.48242289 0.5300493909
#>  [6,] 0.8920778 0.12848083 0.5752964869
#>  [7,] 0.2769754 0.67252348 0.2750262578
#>  [8,] 2.2010562 1.00278230 1.3536030982
#>  [9,] 1.9698049 0.97939573 0.5532939018
#> [10,] 0.5802938 0.52058142 1.0545752632
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12424551 0.12424551 0.12424551
#>  [2,] 0.85623117 0.85623117 0.85623117
#>  [3,] 1.10348478 1.10348478 1.10348478
#>  [4,] 0.21326375 0.21326375 0.21326375
#>  [5,] 0.13686311 0.13686311 0.13686311
#>  [6,] 0.44420831 0.44420831 0.44420831
#>  [7,] 0.46046291 0.46046291 0.46046291
#>  [8,] 0.10616611 0.10616611 0.10616611
#>  [9,] 0.08511751 0.08511751 0.08511751
#> [10,] 0.68685240 0.68685240 0.68685240

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.85181377 0.03905771 1.8518138
#>  [2,] 0.27464375 0.94438748 1.5971705
#>  [3,] 0.47254295 0.47254295 0.4725430
#>  [4,] 0.14580142 0.27832981 0.1458014
#>  [5,] 0.56631504 0.56631504 0.5663150
#>  [6,] 2.26659954 1.78400456 0.4881285
#>  [7,] 0.04448974 0.58787860 0.5878786
#>  [8,] 0.46226634 0.18795514 0.5060004
#>  [9,] 0.18396040 0.24535111 0.4568318
#> [10,] 1.09993232 0.54335571 0.5433557
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23689791 0.47963214 1.82775650
#>  [2,] 0.85005789 0.74413375 1.64993131
#>  [3,] 0.11982137 0.71298025 0.09864844
#>  [4,] 0.19115542 0.06522712 0.23460049
#>  [5,] 0.47184532 0.74038040 2.59842332
#>  [6,] 0.07703903 0.97460178 0.33480661
#>  [7,] 0.39087558 0.64593916 0.17994341
#>  [8,] 1.40140269 1.10801101 0.20813383
#>  [9,] 0.89361119 3.47753601 0.72067451
#> [10,] 2.09453259 2.44707049 0.38090595
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.96253839 0.96253839 0.96253839
#>  [2,] 0.30662462 0.30662462 0.30662462
#>  [3,] 0.09501820 0.09501820 0.09501820
#>  [4,] 3.98113972 3.98113972 3.98113972
#>  [5,] 2.14397809 2.14397809 2.14397809
#>  [6,] 0.63112905 0.63112905 0.63112905
#>  [7,] 0.07714265 0.07714265 0.07714265
#>  [8,] 0.29670498 0.29670498 0.29670498
#>  [9,] 1.16339308 1.16339308 1.16339308
#> [10,] 1.47197221 1.47197221 1.47197221