Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationsip of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5278447 1.3038270 2.9317400
#>  [2,] 2.8453608 0.2321212 1.6517436
#>  [3,] 0.6477666 0.6477666 0.6477666
#>  [4,] 0.4526132 1.9559869 1.9680873
#>  [5,] 0.1365134 0.7688323 0.7688323
#>  [6,] 3.0447432 0.2649740 0.7097131
#>  [7,] 0.3163667 0.3163667 0.3163667
#>  [8,] 0.2925419 0.2925419 0.2925419
#>  [9,] 1.7303818 1.4248834 1.4627690
#> [10,] 0.7805673 1.1083419 0.4777507
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.0894736 1.06171893 0.11295514
#>  [2,] 1.2172095 1.30774742 0.29152482
#>  [3,] 1.0053752 0.59051225 2.17035907
#>  [4,] 1.1649825 0.39388001 0.03353294
#>  [5,] 0.1176509 0.59667378 0.90651636
#>  [6,] 0.4698111 2.50503018 0.01982371
#>  [7,] 0.5386177 0.11384745 1.35655308
#>  [8,] 0.7667029 0.08712008 1.39636131
#>  [9,] 0.4476167 0.86024100 1.29149902
#> [10,] 0.3745100 1.08589779 0.82589400
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7321727 0.7321727 0.7321727
#>  [2,] 0.9443693 0.9443693 0.9443693
#>  [3,] 0.9795036 0.9795036 0.9795036
#>  [4,] 4.0324090 4.0324090 4.0324090
#>  [5,] 0.6287705 0.6287705 0.6287705
#>  [6,] 0.3138699 0.3138699 0.3138699
#>  [7,] 1.7426587 1.7426587 1.7426587
#>  [8,] 0.3979609 0.3979609 0.3979609
#>  [9,] 2.9189460 2.9189460 2.9189460
#> [10,] 1.9772052 1.9772052 1.9772052

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.9929786 2.23696800 2.42885853
#>  [2,] 0.5755734 2.50066983 0.83834585
#>  [3,] 0.5876467 1.54202875 1.17171597
#>  [4,] 2.9312929 0.02171197 2.93129289
#>  [5,] 1.4126873 0.58725093 3.63658476
#>  [6,] 0.5243500 0.52435001 0.52435001
#>  [7,] 0.6823976 0.68239759 0.38809581
#>  [8,] 0.3977791 0.39777906 0.05924669
#>  [9,] 0.2716837 1.33626499 0.27168371
#> [10,] 1.1221808 1.43087287 0.62758841
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.61945789 0.19111808 1.4093044
#>  [2,] 0.37469819 0.71477108 0.1488070
#>  [3,] 0.46684327 0.09753414 2.5598410
#>  [4,] 4.74658686 0.66144701 2.2658356
#>  [5,] 0.28443837 0.13650549 2.0686963
#>  [6,] 2.99294624 0.17172564 0.9045176
#>  [7,] 0.07924003 1.12412688 0.2939242
#>  [8,] 0.50262030 0.66676601 0.5042084
#>  [9,] 3.19765373 1.25906502 0.1336953
#> [10,] 1.13539835 1.24829520 0.8760892
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.040990114 1.040990114 1.040990114
#>  [2,] 1.461930899 1.461930899 1.461930899
#>  [3,] 0.378305519 0.378305519 0.378305519
#>  [4,] 1.486298049 1.486298049 1.486298049
#>  [5,] 2.074194500 2.074194500 2.074194500
#>  [6,] 0.462998284 0.462998284 0.462998284
#>  [7,] 0.002884561 0.002884561 0.002884561
#>  [8,] 0.302060410 0.302060410 0.302060410
#>  [9,] 1.038570952 1.038570952 1.038570952
#> [10,] 0.393042462 0.393042462 0.393042462

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9928029 0.2659414 0.9928029
#>  [2,] 2.1277160 0.1913567 2.2917857
#>  [3,] 1.2191665 1.5062919 1.5062919
#>  [4,] 0.6329833 1.6095137 5.2370581
#>  [5,] 0.4743361 0.4743361 0.4743361
#>  [6,] 0.8118154 0.8118154 1.1906073
#>  [7,] 0.4275614 3.0578017 1.1568902
#>  [8,] 0.8983406 1.4493809 0.7089078
#>  [9,] 0.9051656 1.9926151 0.2276864
#> [10,] 1.4535063 0.6078321 0.6895614
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6191579 1.11841908 0.6190116
#>  [2,] 0.1058352 0.83337381 0.5547557
#>  [3,] 0.1997059 0.06476468 0.8426916
#>  [4,] 0.9128961 4.32635330 0.3166325
#>  [5,] 1.5244441 1.73089524 2.0213820
#>  [6,] 0.6136895 0.36774262 0.6505461
#>  [7,] 0.5004194 2.40877798 0.1958236
#>  [8,] 0.5050250 0.64667772 1.1179408
#>  [9,] 0.7799363 0.35170936 1.3228022
#> [10,] 1.1445735 0.63790373 0.9024268
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41628342 0.41628342 0.41628342
#>  [2,] 0.47527731 0.47527731 0.47527731
#>  [3,] 1.06332718 1.06332718 1.06332718
#>  [4,] 1.57691841 1.57691841 1.57691841
#>  [5,] 2.20884421 2.20884421 2.20884421
#>  [6,] 0.54503160 0.54503160 0.54503160
#>  [7,] 0.20795389 0.20795389 0.20795389
#>  [8,] 1.60972141 1.60972141 1.60972141
#>  [9,] 0.87585175 0.87585175 0.87585175
#> [10,] 0.09227124 0.09227124 0.09227124

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.70876116 0.70876116 0.3664179
#>  [2,] 0.26993917 1.53073632 1.3180651
#>  [3,] 0.03413995 0.03413995 0.9417370
#>  [4,] 0.64698864 0.01026787 4.8657329
#>  [5,] 0.54186375 0.89294615 0.3343731
#>  [6,] 0.06772440 0.06772440 0.1162828
#>  [7,] 0.62616875 0.62616875 0.6261687
#>  [8,] 0.50140652 1.27728460 0.8222066
#>  [9,] 1.32593139 0.26039375 1.1684761
#> [10,] 0.42626563 0.28762905 0.8364851
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.48380669 0.8258582 0.94149691
#>  [2,] 0.17053874 3.0306044 2.91347002
#>  [3,] 0.14514826 0.5700755 0.02487145
#>  [4,] 1.68184069 1.1361813 3.17572983
#>  [5,] 0.42186962 0.1419878 1.67048515
#>  [6,] 0.04884654 0.8541591 1.15263154
#>  [7,] 0.87491076 0.1138047 0.11418411
#>  [8,] 0.59840036 0.3004754 0.57536650
#>  [9,] 0.49283548 0.8169096 0.33932315
#> [10,] 0.26367241 1.8152006 0.30278244
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4251201 0.4251201 0.4251201
#>  [2,] 2.0716901 2.0716901 2.0716901
#>  [3,] 1.4259216 1.4259216 1.4259216
#>  [4,] 3.1447048 3.1447048 3.1447048
#>  [5,] 0.4680890 0.4680890 0.4680890
#>  [6,] 0.4862143 0.4862143 0.4862143
#>  [7,] 3.3388253 3.3388253 3.3388253
#>  [8,] 1.5829265 1.5829265 1.5829265
#>  [9,] 2.5237294 2.5237294 2.5237294
#> [10,] 0.6997772 0.6997772 0.6997772