Sample from exchangeable Marshall–Olkin distributions

Draws n iid samples from a d-variate exchangeable Marshall–Olkin distribution parametrized by exchangeable shock-size-arrival intensities.

Usage

rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

ex_intensities

a numeric vector with the exchangeable shock-size-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rexmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate exchangeable Marshall–Olkin distribution with exchangeable shock-size-arrival intensities

ex_intensities.

Details

The exchangeable Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ -\sum_{i=1}^{d}{ \eta_{i} \tau_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for exchangeable shock-size-arrival intensities \(\eta_{i} \geq 0\), \(1 \leq i \leq d\) and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . They are called shock-size-arrival intensities as they correspond to the rates of the minimums of all independent exponential random variables corresponding to \(I\)-sized shocks from the exogenous shock model (ESM), and a shock-size-arrival intensity \(\eta_{i}\) of shock-size \(i\) equal to zero implies that no shock with size \(i\) occurs. The relationship of exchangeable shock-size-arrival intensities to the shock-arrival intensities of the Marshall–Olkin distribution, see rmo(), is as follows: $$ \eta_{i} = \binom{d}{i} \lambda_{i} . $$

Simulation algorithms

General Marshall–Olkin sampling algorithms

The exogenous shock model (ESM) and Arnold model (AM) simulation algorithms for the general Marshall–Olkin distribution can be used. For this, the exchangeable shock-size-arrival intensities are converted to the corresponding shock-arrival intensities and passed to rmo().

Markovian death-set model

TBD; see (Sloot 2022) and (see pp. 122 psqq. Mai and Scherer 2017) for a similar algorithm.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also

Examples

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9898261 1.5008766 1.5008766
#>  [2,] 0.1828582 0.4395785 0.6032284
#>  [3,] 0.9052887 0.8845479 1.1172014
#>  [4,] 1.6594836 1.5485986 1.6594836
#>  [5,] 0.3328797 0.7524100 0.9565839
#>  [6,] 0.3582865 0.3582865 0.3582865
#>  [7,] 0.5915362 1.1444195 2.5922108
#>  [8,] 0.6895310 0.5533971 2.7986177
#>  [9,] 0.7398036 1.7238748 1.6008994
#> [10,] 2.8336533 2.7668154 2.8336533
## independence
rexmo(
    10, 3,
    c(3, 0, 0)
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.6023053 0.05287297 0.2232123
#>  [2,] 0.7388479 1.55800633 1.0039593
#>  [3,] 0.5801676 1.24022252 0.2736037
#>  [4,] 0.3789085 0.56535201 1.0846014
#>  [5,] 0.4846830 0.01640276 0.9607377
#>  [6,] 0.5519969 2.60479937 0.4014911
#>  [7,] 0.4327788 2.35459223 1.4809263
#>  [8,] 1.5344160 0.32883738 1.9995128
#>  [9,] 0.1413125 0.23003359 0.5481564
#> [10,] 0.3608735 2.92425806 0.8742783
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 5.6341488 5.6341488 5.6341488
#>  [2,] 0.1787573 0.1787573 0.1787573
#>  [3,] 1.1733411 1.1733411 1.1733411
#>  [4,] 1.4245068 1.4245068 1.4245068
#>  [5,] 0.9048049 0.9048049 0.9048049
#>  [6,] 0.8806611 0.8806611 0.8806611
#>  [7,] 1.8441893 1.8441893 1.8441893
#>  [8,] 0.9910664 0.9910664 0.9910664
#>  [9,] 1.9680901 1.9680901 1.9680901
#> [10,] 2.2517103 2.2517103 2.2517103

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38363274 0.38363274 0.38363274
#>  [2,] 1.36341414 1.36341414 0.39172555
#>  [3,] 0.66889129 0.66889129 0.66889129
#>  [4,] 1.90831264 0.78767619 0.08144966
#>  [5,] 1.01102731 0.74344714 1.17785356
#>  [6,] 1.14661449 0.29497996 0.10206652
#>  [7,] 0.06106851 2.57783839 0.43838380
#>  [8,] 0.04147067 0.04147067 0.02189034
#>  [9,] 1.13557463 2.98675121 1.44267262
#> [10,] 2.11756285 2.11756285 2.06140463
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.62969874 0.2238060 0.36224674
#>  [2,] 0.07439974 0.7837034 0.28251470
#>  [3,] 0.40221702 0.7416878 0.41221828
#>  [4,] 0.34596469 0.3276538 0.92503924
#>  [5,] 4.11254600 0.3614542 1.22507098
#>  [6,] 1.84461766 0.2356221 2.05371102
#>  [7,] 0.69798944 0.3602363 0.06391012
#>  [8,] 1.02350499 0.5616695 0.44315349
#>  [9,] 0.52166359 0.5948032 0.17455106
#> [10,] 0.10038260 0.2959227 0.41405056
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9950991 0.9950991 0.9950991
#>  [2,] 1.7307810 1.7307810 1.7307810
#>  [3,] 1.5279716 1.5279716 1.5279716
#>  [4,] 0.1275816 0.1275816 0.1275816
#>  [5,] 0.6514046 0.6514046 0.6514046
#>  [6,] 0.2971093 0.2971093 0.2971093
#>  [7,] 3.9601084 3.9601084 3.9601084
#>  [8,] 0.4556899 0.4556899 0.4556899
#>  [9,] 1.0624531 1.0624531 1.0624531
#> [10,] 0.3481476 0.3481476 0.3481476

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5108867 1.7839522 0.4938798
#>  [2,] 0.1341659 4.8903241 1.2285746
#>  [3,] 0.9671992 3.0853253 3.0853253
#>  [4,] 0.8303414 0.8303414 1.1349499
#>  [5,] 2.7685195 0.2561734 0.3213394
#>  [6,] 0.5790082 0.5790082 0.5790082
#>  [7,] 0.2561100 0.3798131 1.6945054
#>  [8,] 0.5981492 1.0619069 1.3469554
#>  [9,] 0.5123729 0.5123729 0.5123729
#> [10,] 0.4818798 2.7594324 2.9716114
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "AM"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 0.4480737 0.722762386 0.44413871
#>  [2,] 1.8006472 0.971970235 1.19718067
#>  [3,] 1.7112264 0.146259825 0.41420249
#>  [4,] 2.0772722 0.008832213 0.26992440
#>  [5,] 0.4409475 0.518219931 1.51990430
#>  [6,] 2.2443205 0.044342538 0.70451495
#>  [7,] 0.7644449 0.651721391 1.10816538
#>  [8,] 0.3033118 0.767559251 1.15373226
#>  [9,] 0.9922779 1.706087455 0.09343171
#> [10,] 0.3487371 0.060376912 0.57124804
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.62478698 0.62478698 0.62478698
#>  [2,] 0.73012813 0.73012813 0.73012813
#>  [3,] 0.21134897 0.21134897 0.21134897
#>  [4,] 0.10101126 0.10101126 0.10101126
#>  [5,] 0.86410075 0.86410075 0.86410075
#>  [6,] 0.34817357 0.34817357 0.34817357
#>  [7,] 0.15160103 0.15160103 0.15160103
#>  [8,] 1.10973249 1.10973249 1.10973249
#>  [9,] 0.09422372 0.09422372 0.09422372
#> [10,] 0.06307205 0.06307205 0.06307205

rexmo(
    10, 3,
    c(1.2, 0.3, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.05801677 2.16835696 0.60301753
#>  [2,] 2.27686687 2.29785486 3.67671354
#>  [3,] 0.09198043 0.09198043 0.68694517
#>  [4,] 0.94421103 0.17032885 0.01032205
#>  [5,] 2.43587389 0.43657772 2.43587389
#>  [6,] 1.28231493 1.66988901 1.28231493
#>  [7,] 1.34855608 1.34855608 0.12831564
#>  [8,] 0.47586391 0.47586391 0.47586391
#>  [9,] 0.61301958 1.45409701 0.80294027
#> [10,] 0.17579532 0.26570390 0.26570390
## independence
rexmo(
    10, 3,
    c(3, 0, 0),
    method = "ESM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.07373506 0.15778173 1.4916333
#>  [2,] 0.01965232 0.08547246 1.5320913
#>  [3,] 1.17545492 1.33702666 0.5421537
#>  [4,] 0.31819491 1.21012108 1.6065198
#>  [5,] 0.93582124 1.12263604 1.1296161
#>  [6,] 0.45867214 2.26225264 1.6077670
#>  [7,] 0.22923076 0.01216547 0.4643569
#>  [8,] 0.30788134 0.63549713 0.4641524
#>  [9,] 1.02782613 0.78267337 1.0702691
#> [10,] 0.72923488 0.04647891 0.9202040
## comonotone
rexmo(
    10, 3,
    c(0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07767481 0.07767481 0.07767481
#>  [2,] 0.11246158 0.11246158 0.11246158
#>  [3,] 0.12019201 0.12019201 0.12019201
#>  [4,] 0.74392470 0.74392470 0.74392470
#>  [5,] 1.01817297 1.01817297 1.01817297
#>  [6,] 0.27851339 0.27851339 0.27851339
#>  [7,] 0.65947749 0.65947749 0.65947749
#>  [8,] 1.87321120 1.87321120 1.87321120
#>  [9,] 0.37760892 0.37760892 0.37760892
#> [10,] 0.14159380 0.14159380 0.14159380