Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.97159888 0.97159888 2.05468113
#>  [2,] 0.51545241 0.30379447 0.99544785
#>  [3,] 0.12513115 0.12513115 0.12513115
#>  [4,] 0.63242344 1.36911386 1.36911386
#>  [5,] 0.93024212 0.15367293 0.15367293
#>  [6,] 0.12859807 0.29029883 1.00194789
#>  [7,] 1.26018070 2.66641098 1.15557235
#>  [8,] 0.86530425 0.41273415 1.10750124
#>  [9,] 2.14593746 2.46130569 2.46130569
#> [10,] 0.06647178 0.06647178 0.05618354
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23502622 1.68253541 1.30044455
#>  [2,] 0.05179661 0.39974882 0.58874712
#>  [3,] 1.81508122 0.04418315 0.21693909
#>  [4,] 0.30855302 0.67953267 1.43085123
#>  [5,] 3.31883683 4.02461453 0.12102504
#>  [6,] 1.35080261 0.61617866 0.26746667
#>  [7,] 1.78414372 3.21691304 0.41493468
#>  [8,] 0.30902550 0.94294103 2.58313294
#>  [9,] 0.50780586 1.58864394 0.05369671
#> [10,] 1.24118854 0.19920973 0.91281404
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58125706 0.58125706 0.58125706
#>  [2,] 0.70487884 0.70487884 0.70487884
#>  [3,] 0.98771909 0.98771909 0.98771909
#>  [4,] 2.33175120 2.33175120 2.33175120
#>  [5,] 1.90866436 1.90866436 1.90866436
#>  [6,] 0.55436049 0.55436049 0.55436049
#>  [7,] 0.92089429 0.92089429 0.92089429
#>  [8,] 0.03836483 0.03836483 0.03836483
#>  [9,] 1.25902533 1.25902533 1.25902533
#> [10,] 3.47624613 3.47624613 3.47624613

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.64905372 0.64905372 0.64905372
#>  [2,] 0.04542018 0.87182121 0.87182121
#>  [3,] 0.76106567 0.76106567 0.66661115
#>  [4,] 1.48978789 1.48978789 0.16158810
#>  [5,] 1.55018614 0.46674214 2.03320857
#>  [6,] 0.09572936 0.84693407 0.84693407
#>  [7,] 0.41171687 0.41171687 0.41171687
#>  [8,] 0.63859152 1.71260969 2.22717662
#>  [9,] 0.06135082 0.06135082 0.06135082
#> [10,] 0.05768221 0.59923062 0.59923062
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>            [,1]      [,2]        [,3]
#>  [1,] 4.1175000 1.6820939 0.186214169
#>  [2,] 1.3268408 1.0723297 0.491923523
#>  [3,] 0.4922257 0.6400083 1.089557432
#>  [4,] 0.7375757 1.1840735 0.828739705
#>  [5,] 2.0217019 1.1909418 1.340915763
#>  [6,] 1.4501909 0.3362487 0.868691696
#>  [7,] 1.2666519 1.1071382 0.196715208
#>  [8,] 0.5594288 1.0325531 1.327007591
#>  [9,] 0.1612112 0.7352041 0.005392984
#> [10,] 1.3960485 1.3182501 1.013974704
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.45653239 0.45653239 0.45653239
#>  [2,] 0.21619940 0.21619940 0.21619940
#>  [3,] 2.09318345 2.09318345 2.09318345
#>  [4,] 2.06795228 2.06795228 2.06795228
#>  [5,] 3.09593604 3.09593604 3.09593604
#>  [6,] 0.57937587 0.57937587 0.57937587
#>  [7,] 0.37661087 0.37661087 0.37661087
#>  [8,] 0.05601293 0.05601293 0.05601293
#>  [9,] 0.08022823 0.08022823 0.08022823
#> [10,] 0.99848436 0.99848436 0.99848436

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.97283312 4.16723285 1.30574770
#>  [2,] 0.32426742 0.39398949 0.32426742
#>  [3,] 0.25948191 0.25948191 0.25948191
#>  [4,] 1.06608431 1.52700735 1.52700735
#>  [5,] 1.34479348 1.34479348 0.33137347
#>  [6,] 0.03439812 0.03439812 0.03439812
#>  [7,] 2.08359691 2.08359691 2.08359691
#>  [8,] 2.30528913 2.40155745 1.95437515
#>  [9,] 3.74068518 2.80697070 1.95771122
#> [10,] 0.08514295 0.80112466 0.53385976
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.05036329 0.1622470 1.76596302
#>  [2,] 3.83614304 0.9598871 1.31758957
#>  [3,] 0.55559366 1.7868107 1.10907924
#>  [4,] 0.85494933 1.4990376 0.08427666
#>  [5,] 0.05731395 0.4064613 2.19295849
#>  [6,] 1.01728064 1.3081919 0.95711270
#>  [7,] 1.21564967 1.9248860 0.40108160
#>  [8,] 0.04119048 0.3242193 1.29436807
#>  [9,] 0.88927140 0.5402999 0.34795315
#> [10,] 1.95370173 0.7487631 0.41606207
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 4.40028026 4.40028026 4.40028026
#>  [2,] 0.64722187 0.64722187 0.64722187
#>  [3,] 0.30032196 0.30032196 0.30032196
#>  [4,] 1.67905662 1.67905662 1.67905662
#>  [5,] 0.45425345 0.45425345 0.45425345
#>  [6,] 0.50715370 0.50715370 0.50715370
#>  [7,] 2.13041864 2.13041864 2.13041864
#>  [8,] 0.08822577 0.08822577 0.08822577
#>  [9,] 0.01143277 0.01143277 0.01143277
#> [10,] 1.83216052 1.83216052 1.83216052