Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.49088995 0.84494881 0.84494881
#>  [2,] 0.03559829 4.35310320 0.03559829
#>  [3,] 3.02692570 3.02692570 0.10186811
#>  [4,] 0.12112621 0.12112621 0.12112621
#>  [5,] 1.03486384 0.12067514 1.18795159
#>  [6,] 1.80844481 1.80844481 0.62120481
#>  [7,] 0.08694575 0.08694575 0.08694575
#>  [8,] 1.17779816 1.17779816 1.17779816
#>  [9,] 1.40090996 1.79692096 0.05880975
#> [10,] 5.19764448 1.21688744 1.65049739
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.2545866 1.33655156 0.30810524
#>  [2,] 0.6491764 1.02778994 1.00402121
#>  [3,] 0.6423782 0.35475287 1.23765850
#>  [4,] 0.7240736 0.78629557 0.91498290
#>  [5,] 0.2217270 0.63238916 1.42364761
#>  [6,] 0.6288405 0.03882245 0.05176723
#>  [7,] 1.7508904 0.39775844 1.24721370
#>  [8,] 0.9236380 2.12743600 0.34649601
#>  [9,] 0.3385339 1.34899794 0.72557738
#> [10,] 0.4801778 1.05473726 0.44083105
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.32468737 1.32468737 1.32468737
#>  [2,] 1.41918888 1.41918888 1.41918888
#>  [3,] 0.40995506 0.40995506 0.40995506
#>  [4,] 0.38493666 0.38493666 0.38493666
#>  [5,] 0.92493079 0.92493079 0.92493079
#>  [6,] 1.65258280 1.65258280 1.65258280
#>  [7,] 1.42725814 1.42725814 1.42725814
#>  [8,] 0.07651246 0.07651246 0.07651246
#>  [9,] 0.28795751 0.28795751 0.28795751
#> [10,] 0.22751116 0.22751116 0.22751116

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1893401 0.79510999 0.79510999
#>  [2,] 0.1539216 0.80878754 0.44205893
#>  [3,] 1.0220310 0.28026084 0.28026084
#>  [4,] 1.7692532 2.91281858 1.76925324
#>  [5,] 0.5150829 0.51508291 2.54835154
#>  [6,] 0.4036220 0.40362199 0.40362199
#>  [7,] 0.2881814 0.28818135 0.28818135
#>  [8,] 0.1773560 0.90087626 1.38873224
#>  [9,] 0.6329423 0.02708837 0.04690119
#> [10,] 1.7392848 0.87486480 1.73928475
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.25258329 3.17409843 1.58901122
#>  [2,] 0.09875860 2.54410441 0.23433017
#>  [3,] 1.83753942 0.60173697 1.44859901
#>  [4,] 1.34148657 0.12642986 1.94936745
#>  [5,] 1.35058526 1.01624510 0.28363733
#>  [6,] 0.36288297 0.18764018 4.76536957
#>  [7,] 0.48419336 0.07525785 0.24926394
#>  [8,] 0.03728496 0.97609241 0.41714146
#>  [9,] 0.41920470 0.81997544 0.05790647
#> [10,] 1.52711196 0.37398734 0.05595814
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4177781 0.4177781 0.4177781
#>  [2,] 0.2592726 0.2592726 0.2592726
#>  [3,] 4.1651578 4.1651578 4.1651578
#>  [4,] 0.9619435 0.9619435 0.9619435
#>  [5,] 0.2834200 0.2834200 0.2834200
#>  [6,] 1.3492012 1.3492012 1.3492012
#>  [7,] 0.6458271 0.6458271 0.6458271
#>  [8,] 0.3050891 0.3050891 0.3050891
#>  [9,] 0.2987203 0.2987203 0.2987203
#> [10,] 0.1020508 0.1020508 0.1020508

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2427524 0.2427524 0.1388711
#>  [2,] 2.0990708 2.2164101 1.8646618
#>  [3,] 0.1335025 0.1335025 0.1335025
#>  [4,] 0.1581173 2.2113154 0.1581173
#>  [5,] 0.1236007 0.5503341 0.2643357
#>  [6,] 1.8379516 1.3562804 1.5183484
#>  [7,] 0.3365411 0.8134949 0.8134949
#>  [8,] 0.9766926 0.8225452 0.8225452
#>  [9,] 2.3714116 2.3714116 1.2643340
#> [10,] 2.2044556 2.2044556 1.3759678
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5572605 0.1987273 0.5437136
#>  [2,] 1.5818660 1.0968354 0.1064394
#>  [3,] 1.5518186 1.7165699 0.2117509
#>  [4,] 0.8400925 0.2079901 0.5561449
#>  [5,] 0.5968357 1.1025826 1.0887880
#>  [6,] 2.1994336 1.0937674 1.7162819
#>  [7,] 3.7372425 1.4589819 0.4432989
#>  [8,] 0.1424775 0.5076218 2.5156489
#>  [9,] 0.2700750 0.9914013 0.9155109
#> [10,] 0.4321612 1.4955738 3.3905431
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13218136 0.13218136 0.13218136
#>  [2,] 0.01076151 0.01076151 0.01076151
#>  [3,] 0.59074488 0.59074488 0.59074488
#>  [4,] 0.32520785 0.32520785 0.32520785
#>  [5,] 3.79662279 3.79662279 3.79662279
#>  [6,] 0.11505073 0.11505073 0.11505073
#>  [7,] 2.06011615 2.06011615 2.06011615
#>  [8,] 0.42176220 0.42176220 0.42176220
#>  [9,] 3.37996908 3.37996908 3.37996908
#> [10,] 1.62812491 1.62812491 1.62812491