Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39887906 0.39887906 0.39887906
#>  [2,] 4.81336479 4.81336479 1.09788472
#>  [3,] 2.46319998 0.07224387 0.76720436
#>  [4,] 0.49262583 2.16962214 1.12669250
#>  [5,] 0.78140414 1.26566893 2.40697528
#>  [6,] 0.79150648 0.79150648 0.33497094
#>  [7,] 0.74389831 0.74389831 1.27805224
#>  [8,] 0.40540463 0.40540463 0.40540463
#>  [9,] 0.00796966 0.00796966 0.00796966
#> [10,] 1.45422665 0.28450907 1.45422665
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]        [,2]        [,3]
#>  [1,] 0.19974614 0.063910339 1.269930106
#>  [2,] 0.54313576 0.249605176 2.150106906
#>  [3,] 0.48464519 1.009852286 0.434621097
#>  [4,] 0.45437664 0.009964604 1.592743727
#>  [5,] 0.28165965 1.321695776 0.007282119
#>  [6,] 0.26538027 0.360969002 0.081587346
#>  [7,] 1.53962553 1.513317327 0.770230551
#>  [8,] 0.05907391 1.231771338 2.607117688
#>  [9,] 0.05638427 0.280248773 0.719209586
#> [10,] 0.69490430 0.486062943 2.392284717
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.71320316 0.71320316 0.71320316
#>  [2,] 0.15820351 0.15820351 0.15820351
#>  [3,] 0.87943001 0.87943001 0.87943001
#>  [4,] 1.30516923 1.30516923 1.30516923
#>  [5,] 1.20087182 1.20087182 1.20087182
#>  [6,] 2.95388932 2.95388932 2.95388932
#>  [7,] 0.12889905 0.12889905 0.12889905
#>  [8,] 4.17671443 4.17671443 4.17671443
#>  [9,] 0.05830124 0.05830124 0.05830124
#> [10,] 0.50291292 0.50291292 0.50291292

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.33461336 0.3346134 0.3346134
#>  [2,] 0.03936278 0.1341975 0.1341975
#>  [3,] 0.46560468 0.2167827 0.4656047
#>  [4,] 0.20093549 0.2009355 0.2009355
#>  [5,] 0.93553731 0.7435987 0.2975883
#>  [6,] 0.27653064 0.6064670 0.2097764
#>  [7,] 0.19138265 1.2520763 0.5989839
#>  [8,] 0.52359450 0.5235945 0.5235945
#>  [9,] 0.84561202 1.3975246 1.3975246
#> [10,] 0.51573824 0.1778438 1.8703093
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.09938906 0.05525110 2.5623624
#>  [2,] 0.46239680 0.04727912 0.8591452
#>  [3,] 0.45762827 1.79885689 2.7894132
#>  [4,] 0.91451779 0.38739285 0.3327567
#>  [5,] 2.66285598 0.12270466 1.2487661
#>  [6,] 0.11288817 4.70199364 0.4109877
#>  [7,] 0.09314827 1.15580560 0.6027251
#>  [8,] 2.51232763 0.27130815 3.9325523
#>  [9,] 1.19734410 0.72593680 1.2027396
#> [10,] 3.80216461 3.62124863 1.8789986
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.75434093 0.75434093 0.75434093
#>  [2,] 1.23181496 1.23181496 1.23181496
#>  [3,] 0.96718122 0.96718122 0.96718122
#>  [4,] 0.43034525 0.43034525 0.43034525
#>  [5,] 1.00390634 1.00390634 1.00390634
#>  [6,] 3.26106118 3.26106118 3.26106118
#>  [7,] 0.27291275 0.27291275 0.27291275
#>  [8,] 3.27026736 3.27026736 3.27026736
#>  [9,] 1.56589799 1.56589799 1.56589799
#> [10,] 0.03529881 0.03529881 0.03529881

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.03448084 0.034480839 0.03448084
#>  [2,] 0.42904573 0.429045734 0.42904573
#>  [3,] 0.56509349 0.005944014 1.22026966
#>  [4,] 0.01590367 0.015903668 0.01590367
#>  [5,] 0.35324116 1.056879480 0.76617654
#>  [6,] 0.74412134 1.488382007 0.74412134
#>  [7,] 0.46332403 0.463324033 0.46332403
#>  [8,] 1.13703821 1.110317995 1.13703821
#>  [9,] 0.73412953 1.067596701 1.06759670
#> [10,] 0.56431053 2.822789481 0.56431053
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6093375 1.5089830 0.1071284
#>  [2,] 0.3434087 1.2304371 1.7958222
#>  [3,] 1.8114723 0.6805754 1.6438646
#>  [4,] 0.5703357 1.7020114 0.2257763
#>  [5,] 0.3185843 0.5343961 1.6421041
#>  [6,] 2.9438706 1.5341349 0.4970302
#>  [7,] 1.1235996 2.2204528 0.2518558
#>  [8,] 1.1004374 0.0148111 0.3427879
#>  [9,] 0.2623702 0.2488779 0.1737259
#> [10,] 1.4816114 0.4353982 2.4025627
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3573731 0.3573731 0.3573731
#>  [2,] 0.1047581 0.1047581 0.1047581
#>  [3,] 1.0317727 1.0317727 1.0317727
#>  [4,] 0.5420995 0.5420995 0.5420995
#>  [5,] 0.5250788 0.5250788 0.5250788
#>  [6,] 0.6995006 0.6995006 0.6995006
#>  [7,] 0.6066957 0.6066957 0.6066957
#>  [8,] 0.2143947 0.2143947 0.2143947
#>  [9,] 0.6467923 0.6467923 0.6467923
#> [10,] 0.8467633 0.8467633 0.8467633