Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>             [,1]      [,2]       [,3]
#>  [1,] 1.89833124 0.1222428 0.12224282
#>  [2,] 0.24617165 0.2461716 0.24617165
#>  [3,] 0.35859053 0.1254463 0.35859053
#>  [4,] 0.37340027 1.4026029 0.98883327
#>  [5,] 0.26017289 0.9191428 0.91914281
#>  [6,] 0.24420638 0.6351183 0.24420638
#>  [7,] 0.58120988 0.5812099 0.58120988
#>  [8,] 0.14906945 0.1490694 0.14906945
#>  [9,] 2.92603159 0.4162058 1.36493808
#> [10,] 0.01947506 1.4780189 0.01947506
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.09969062 4.27598597 0.3108360
#>  [2,] 0.32287998 2.23404926 1.0744220
#>  [3,] 0.46616565 0.01090359 0.7046830
#>  [4,] 0.48266595 0.56224331 0.6068577
#>  [5,] 0.84275236 0.44930030 0.2276872
#>  [6,] 1.66861489 1.35821490 2.4702726
#>  [7,] 0.15222837 2.57154067 0.4971121
#>  [8,] 0.36752661 0.04657145 3.4667442
#>  [9,] 2.30753039 0.87409261 0.3306274
#> [10,] 0.35281101 1.95016550 1.9942634
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6122106 0.6122106 0.6122106
#>  [2,] 4.2983905 4.2983905 4.2983905
#>  [3,] 0.3559268 0.3559268 0.3559268
#>  [4,] 1.0455225 1.0455225 1.0455225
#>  [5,] 0.2246879 0.2246879 0.2246879
#>  [6,] 1.9867289 1.9867289 1.9867289
#>  [7,] 0.1021990 0.1021990 0.1021990
#>  [8,] 0.5604048 0.5604048 0.5604048
#>  [9,] 0.2716739 0.2716739 0.2716739
#> [10,] 0.1766710 0.1766710 0.1766710

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 1.401856577 0.60553467 0.42829674
#>  [2,] 0.007707902 0.04082362 0.04082362
#>  [3,] 1.733917543 1.58867983 1.25439716
#>  [4,] 1.087850004 2.21535192 0.40478355
#>  [5,] 3.599887220 1.41761644 2.15673000
#>  [6,] 0.111126275 0.11112627 1.97677938
#>  [7,] 0.025237932 0.02523793 0.02523793
#>  [8,] 1.872171628 1.50741664 0.14764602
#>  [9,] 0.955516874 1.84767842 0.95551687
#> [10,] 1.146707791 1.14670779 0.30859986
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.02814244 1.21711410 4.1473683
#>  [2,] 0.05166184 0.96628771 1.4720653
#>  [3,] 0.82941871 0.93575205 1.2259197
#>  [4,] 1.08443047 1.29863165 1.2214054
#>  [5,] 1.55846284 0.89234383 1.1427941
#>  [6,] 0.72187751 0.04687898 1.8824435
#>  [7,] 0.44448087 0.81347270 2.6224501
#>  [8,] 0.08291920 0.72456653 1.8618927
#>  [9,] 0.45729465 0.40804402 0.0077733
#> [10,] 1.02710206 0.65756604 0.3940760
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.17857578 2.17857578 2.17857578
#>  [2,] 0.39637418 0.39637418 0.39637418
#>  [3,] 0.46809287 0.46809287 0.46809287
#>  [4,] 0.14421743 0.14421743 0.14421743
#>  [5,] 1.04304549 1.04304549 1.04304549
#>  [6,] 2.45149871 2.45149871 2.45149871
#>  [7,] 1.41782549 1.41782549 1.41782549
#>  [8,] 0.05848195 0.05848195 0.05848195
#>  [9,] 1.92147977 1.92147977 1.92147977
#> [10,] 0.89045642 0.89045642 0.89045642

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.93682327 2.09124113 0.94428660
#>  [2,] 1.59832150 2.00877526 0.59088283
#>  [3,] 0.37792804 0.06006850 0.37792804
#>  [4,] 0.45598052 0.45598052 0.45598052
#>  [5,] 0.01532995 0.01532995 0.01532995
#>  [6,] 1.57798908 1.57798908 1.26085867
#>  [7,] 0.07846356 2.48348978 2.48348978
#>  [8,] 0.96743166 0.96743166 0.05804918
#>  [9,] 0.12017431 0.86304380 2.98766638
#> [10,] 0.79084792 1.98602736 1.98602736
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]       [,2]         [,3]
#>  [1,] 1.0701127 0.19702113 0.0971930624
#>  [2,] 2.1960622 0.05438459 1.7110333349
#>  [3,] 0.6670075 0.47330155 0.0002111313
#>  [4,] 2.1526626 1.33647251 1.0464262524
#>  [5,] 1.0682242 0.08560097 1.5499926488
#>  [6,] 0.2701990 1.45110993 0.0887538720
#>  [7,] 0.5639350 0.01698405 0.3896229678
#>  [8,] 0.7500354 0.41112819 0.6544357580
#>  [9,] 0.1154105 3.94400625 4.3465527496
#> [10,] 0.7636514 0.44680389 4.0387189144
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1720545 0.1720545 0.1720545
#>  [2,] 0.5573917 0.5573917 0.5573917
#>  [3,] 0.8845955 0.8845955 0.8845955
#>  [4,] 0.1107866 0.1107866 0.1107866
#>  [5,] 0.7140576 0.7140576 0.7140576
#>  [6,] 0.2670792 0.2670792 0.2670792
#>  [7,] 1.7196588 1.7196588 1.7196588
#>  [8,] 1.1486041 1.1486041 1.1486041
#>  [9,] 0.3590572 0.3590572 0.3590572
#> [10,] 0.4697706 0.4697706 0.4697706