Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

Calculation of the shock-size arrival intensities

The exchangeable shock-size arrival intensities of a constant Bernstein function with triplet \((a, 0, 0)\) are $$ \eta_{i} = \binom{d}{i} a 1_{\{ i = d \}} , \quad i \in {\{ 1 , \ldots , d \}} . $$

The exchangeable shock-size arrival intensities of a linear Bernstein function with triplet \((0, b, 0)\) are $$ \eta_{i} = \binom{d}{i} b 1_{\{ i = 1 \}} , \quad i \in {\{ 1 , \ldots , d \}} . $$

The exchangeable shock-size arrival intensities of a Bernstein function with Lévy triple \((0, 0, \nu)\) are $$ \eta_{i} = \int_{0}^{\infty} \binom{d}{i} {(1 - e^{-u})}^{i} e^{- u {(d-i)}} \nu{(du)} , \quad i \in {\{ 1, \ldots, d \}} $$

The exchangeable shock-size arrival intensities of a complete Bernstein function with Stieltjes triple \((0, 0, \sigma)\) are $$ \eta_{i} = \int_{0}^{\infty} \binom{d}{i} u B{(1 + i, d - i + u)} \sigma{(du)} , \quad i \in {\{ 1, \ldots, d \}} $$

The exchangeable shock-size arrival intensities of convex combinations of Bernstein functions are the corresponding convex combinations of their individual exchangeable shock-size arrival intensities with the same coefficients.

See (Sloot 2022) for the background, and detailed notes on numerical approximations of these integrals.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.480946913 1.484417760 0.244389008
#>  [2,] 0.253654458 0.253654458 0.253654458
#>  [3,] 0.007095965 0.007095965 0.007095965
#>  [4,] 0.078034758 0.190206151 0.190206151
#>  [5,] 0.064828504 0.062472257 0.064828504
#>  [6,] 0.621072383 0.209882077 0.298677798
#>  [7,] 0.041557921 0.041557921 0.041557921
#>  [8,] 0.313163359 0.399550691 5.100232693
#>  [9,] 0.005877562 1.194367921 0.547513123
#> [10,] 1.642491486 1.642491486 1.642491486
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.49947214 1.77365949 2.8447767
#>  [2,] 2.40780621 1.39008762 2.6064237
#>  [3,] 0.89273574 0.73327965 0.6413030
#>  [4,] 1.96642393 1.72823351 0.2358180
#>  [5,] 0.92052410 1.27764894 0.7622456
#>  [6,] 0.87477977 0.07017228 0.8921969
#>  [7,] 3.08478691 4.85440276 2.1023059
#>  [8,] 0.06866971 0.77324121 0.2349925
#>  [9,] 1.35229594 0.27127949 0.1510841
#> [10,] 2.47106911 0.09527768 0.6761012
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8973087 0.8973087 0.8973087
#>  [2,] 0.1787292 0.1787292 0.1787292
#>  [3,] 0.6834563 0.6834563 0.6834563
#>  [4,] 0.7322801 0.7322801 0.7322801
#>  [5,] 1.2457798 1.2457798 1.2457798
#>  [6,] 0.1530485 0.1530485 0.1530485
#>  [7,] 1.3871418 1.3871418 1.3871418
#>  [8,] 0.8019436 0.8019436 0.8019436
#>  [9,] 0.7190201 0.7190201 0.7190201
#> [10,] 1.0757456 1.0757456 1.0757456

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>              [,1]       [,2]      [,3]
#>  [1,] 0.004918032 2.44681358 0.7328102
#>  [2,] 1.342462061 0.56930333 1.3424621
#>  [3,] 0.395176579 0.89196188 0.3951766
#>  [4,] 1.067693046 1.06769305 1.0676930
#>  [5,] 1.160756464 1.32666118 1.3266612
#>  [6,] 0.369625458 2.33784399 0.1395416
#>  [7,] 0.810997348 0.20737307 0.8109973
#>  [8,] 4.312418388 1.83361047 4.3124184
#>  [9,] 0.962731277 0.45023363 0.4502336
#> [10,] 4.820566413 0.00788046 0.4967425
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.30554019 0.1226033 0.07736523
#>  [2,] 1.88055662 3.5150938 2.40503803
#>  [3,] 0.17435971 0.5256311 0.75508524
#>  [4,] 0.74003054 0.7106616 0.78902214
#>  [5,] 0.08068826 4.6250196 1.25642060
#>  [6,] 1.68363102 0.6084047 0.46031254
#>  [7,] 1.44674169 1.1165195 0.16599340
#>  [8,] 0.05984449 0.3090182 0.48868404
#>  [9,] 0.73773264 1.8502230 2.88212365
#> [10,] 0.86418942 1.7572741 1.07123083
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.35846686 1.35846686 1.35846686
#>  [2,] 0.53158465 0.53158465 0.53158465
#>  [3,] 0.27983174 0.27983174 0.27983174
#>  [4,] 1.44240225 1.44240225 1.44240225
#>  [5,] 1.15119486 1.15119486 1.15119486
#>  [6,] 0.12552260 0.12552260 0.12552260
#>  [7,] 0.06216812 0.06216812 0.06216812
#>  [8,] 1.66860484 1.66860484 1.66860484
#>  [9,] 1.60013687 1.60013687 1.60013687
#> [10,] 1.20546888 1.20546888 1.20546888

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 2.66358668 3.317577021 1.46633328
#>  [2,] 2.99135563 0.002730874 0.99370816
#>  [3,] 2.11067503 0.424896262 0.22337882
#>  [4,] 0.81593244 0.815932437 0.81593244
#>  [5,] 1.17976286 1.388099837 0.06107676
#>  [6,] 0.14773289 0.340075872 0.14773289
#>  [7,] 2.01063369 1.172860222 0.06100279
#>  [8,] 0.47428549 0.066565997 0.47428549
#>  [9,] 0.06521948 0.065219481 0.48888292
#> [10,] 0.67347159 0.673471592 0.67347159
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.8386691 0.4492411 2.14592004
#>  [2,] 1.1179297 0.5985452 0.09004969
#>  [3,] 1.1786282 0.8401248 1.73788984
#>  [4,] 2.2657276 0.8265992 0.80971258
#>  [5,] 0.7936806 0.8631191 0.42196616
#>  [6,] 0.5258533 0.5978867 0.77121299
#>  [7,] 0.4496556 1.1371584 1.68190191
#>  [8,] 0.1932754 0.9673621 0.05233152
#>  [9,] 0.3559045 2.0629498 0.23855594
#> [10,] 0.4834195 0.3050321 0.68372773
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1931194 1.1931194 1.1931194
#>  [2,] 1.3025325 1.3025325 1.3025325
#>  [3,] 3.2475602 3.2475602 3.2475602
#>  [4,] 1.8421558 1.8421558 1.8421558
#>  [5,] 1.1400933 1.1400933 1.1400933
#>  [6,] 2.9330906 2.9330906 2.9330906
#>  [7,] 0.2915521 0.2915521 0.2915521
#>  [8,] 3.1170682 3.1170682 3.1170682
#>  [9,] 0.3577789 0.3577789 0.3577789
#> [10,] 0.9308644 0.9308644 0.9308644