Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 5.47173251 0.46742330 1.07435163
#>  [2,] 1.94965654 1.02609416 0.72691798
#>  [3,] 0.05568804 0.05568804 0.05568804
#>  [4,] 0.76193971 0.03777465 0.12869605
#>  [5,] 0.74692675 0.02152120 0.80516101
#>  [6,] 0.23663067 1.46023004 1.46023004
#>  [7,] 0.57301330 0.44439447 0.57301330
#>  [8,] 1.90395134 0.34269904 1.32331381
#>  [9,] 0.11007219 0.84312530 0.55620473
#> [10,] 2.56167795 1.50574820 4.37626533
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.55675506 1.5930720 0.14813068
#>  [2,] 0.72532860 0.1172535 3.17839577
#>  [3,] 0.02215595 1.5336565 0.57695635
#>  [4,] 2.33806850 0.2073696 0.56745752
#>  [5,] 0.58402502 4.4520095 0.31267399
#>  [6,] 0.78043894 2.3433702 0.75763677
#>  [7,] 0.36324026 0.1084073 0.37047076
#>  [8,] 0.18297058 0.8434863 1.46048599
#>  [9,] 0.40804685 1.3613476 0.57442590
#> [10,] 2.35903440 0.9397736 0.06418591
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.26799586 1.26799586 1.26799586
#>  [2,] 0.39391753 0.39391753 0.39391753
#>  [3,] 3.56971753 3.56971753 3.56971753
#>  [4,] 1.35143932 1.35143932 1.35143932
#>  [5,] 2.70669944 2.70669944 2.70669944
#>  [6,] 1.71383929 1.71383929 1.71383929
#>  [7,] 0.80410059 0.80410059 0.80410059
#>  [8,] 0.06084201 0.06084201 0.06084201
#>  [9,] 0.21923091 0.21923091 0.21923091
#> [10,] 2.37551433 2.37551433 2.37551433

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1557941 0.1557941 0.15579412
#>  [2,] 0.3683064 2.0578640 0.77274850
#>  [3,] 1.3145642 1.6326898 1.40463890
#>  [4,] 0.2415995 0.2415995 0.24159946
#>  [5,] 0.1234776 0.1234776 0.02141373
#>  [6,] 1.8751247 1.8751247 1.51966354
#>  [7,] 0.5518082 0.5518082 0.55180824
#>  [8,] 1.4468463 1.4450722 2.64293043
#>  [9,] 1.8200426 2.0206948 0.35986535
#> [10,] 0.4622155 0.4622155 0.52541258
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.86489559 0.28523926 0.24304397
#>  [2,] 0.78388055 0.32336822 0.05103751
#>  [3,] 0.18661036 0.06136597 1.83728470
#>  [4,] 4.97437200 1.07358080 1.40693868
#>  [5,] 1.80863091 1.09736224 0.22894956
#>  [6,] 2.52070534 0.12260009 1.14746948
#>  [7,] 0.33553068 0.81910667 0.34175550
#>  [8,] 0.06729856 2.34984821 0.15365743
#>  [9,] 1.06950341 1.83592365 1.07504039
#> [10,] 0.97556037 0.31151587 0.73099071
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.99964882 2.99964882 2.99964882
#>  [2,] 1.89552826 1.89552826 1.89552826
#>  [3,] 0.72946891 0.72946891 0.72946891
#>  [4,] 0.41797761 0.41797761 0.41797761
#>  [5,] 1.28555737 1.28555737 1.28555737
#>  [6,] 0.69013684 0.69013684 0.69013684
#>  [7,] 0.92366752 0.92366752 0.92366752
#>  [8,] 1.88560805 1.88560805 1.88560805
#>  [9,] 0.04310646 0.04310646 0.04310646
#> [10,] 1.42676298 1.42676298 1.42676298

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.8444338 1.45932745 1.84443378
#>  [2,] 2.5592889 2.55928889 2.10386550
#>  [3,] 1.0260472 1.30785254 0.76415894
#>  [4,] 1.3805884 1.59311883 1.59311883
#>  [5,] 4.3193229 1.18843381 0.37791821
#>  [6,] 2.2547254 0.03478725 0.03478725
#>  [7,] 0.6113152 0.08269098 0.36492844
#>  [8,] 1.9176827 1.94996766 0.81540766
#>  [9,] 0.1837303 2.57334991 0.20434441
#> [10,] 1.5063582 2.09605954 2.09605954
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.37014385 0.1240311 1.4952845
#>  [2,] 1.02568636 2.1787597 1.1189990
#>  [3,] 1.38090052 0.3476246 1.4817608
#>  [4,] 0.74737507 0.3178397 0.3948638
#>  [5,] 0.84108767 0.1630231 2.6284675
#>  [6,] 1.26459562 0.1790252 0.4817145
#>  [7,] 0.06412288 1.3193748 0.0389041
#>  [8,] 0.28763671 0.1664391 2.6789979
#>  [9,] 0.10973944 0.4170070 2.6683346
#> [10,] 0.63398687 1.4963088 0.3870765
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5414147 0.5414147 0.5414147
#>  [2,] 0.2315174 0.2315174 0.2315174
#>  [3,] 0.3398563 0.3398563 0.3398563
#>  [4,] 0.3776161 0.3776161 0.3776161
#>  [5,] 0.4817125 0.4817125 0.4817125
#>  [6,] 0.1710660 0.1710660 0.1710660
#>  [7,] 0.1158193 0.1158193 0.1158193
#>  [8,] 0.5967429 0.5967429 0.5967429
#>  [9,] 0.5023643 0.5023643 0.5023643
#> [10,] 0.6859712 0.6859712 0.6859712