Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23015369 0.17295324 0.23015369
#>  [2,] 2.37989357 1.54245975 2.37989357
#>  [3,] 0.45698521 0.09030609 0.55715623
#>  [4,] 0.44313746 0.44313746 1.31008006
#>  [5,] 0.22824807 0.22824807 0.02206265
#>  [6,] 0.05121424 0.20001927 0.20001927
#>  [7,] 0.84063279 0.53350659 0.32305842
#>  [8,] 0.92887688 0.92887688 0.92887688
#>  [9,] 1.72041729 0.48093794 0.46539055
#> [10,] 1.19924135 1.19924135 1.19924135
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]         [,3]
#>  [1,] 0.75448068 4.17803822 2.2832656619
#>  [2,] 1.16046426 0.60769991 0.6888038329
#>  [3,] 1.55164468 0.92451242 0.3716210099
#>  [4,] 0.40989091 0.02328100 0.0569328211
#>  [5,] 3.08642726 5.38634410 0.3847614527
#>  [6,] 2.09590831 0.75060540 1.1760114182
#>  [7,] 0.05734079 0.08800627 0.0006589134
#>  [8,] 0.72897358 0.50836998 0.3511053528
#>  [9,] 0.02074194 0.11582452 0.6438545331
#> [10,] 1.24241181 1.88460147 1.3749424243
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.3716975 2.3716975 2.3716975
#>  [2,] 0.3410093 0.3410093 0.3410093
#>  [3,] 0.1761344 0.1761344 0.1761344
#>  [4,] 0.3912159 0.3912159 0.3912159
#>  [5,] 0.1364389 0.1364389 0.1364389
#>  [6,] 0.1970784 0.1970784 0.1970784
#>  [7,] 0.6608344 0.6608344 0.6608344
#>  [8,] 0.1186048 0.1186048 0.1186048
#>  [9,] 1.2936292 1.2936292 1.2936292
#> [10,] 0.9877115 0.9877115 0.9877115

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.16540588 0.47549194 0.47549194
#>  [2,] 0.53826993 0.53826993 0.53826993
#>  [3,] 0.39355117 1.13760565 1.13760565
#>  [4,] 1.09338984 1.01795267 1.09338984
#>  [5,] 1.02309567 1.02309567 0.02397388
#>  [6,] 0.30270142 0.30270142 0.30270142
#>  [7,] 0.06228438 0.06228438 0.20209171
#>  [8,] 0.66522284 1.74358465 1.74358465
#>  [9,] 5.70315529 5.31708089 1.31284142
#> [10,] 0.37375044 1.15339867 0.29167284
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21906116 0.01184090 0.10184228
#>  [2,] 0.26506619 2.65476072 0.63837946
#>  [3,] 0.61640386 0.92964685 0.51671677
#>  [4,] 0.09356167 0.73138068 0.37115298
#>  [5,] 0.76305300 1.25990407 0.41257911
#>  [6,] 0.98158923 1.26002097 0.02448164
#>  [7,] 2.52903714 0.07785733 0.09893347
#>  [8,] 0.61231433 2.89017921 0.25670747
#>  [9,] 1.73417564 1.42725752 2.20535974
#> [10,] 0.53893393 1.28062677 0.74865760
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06792259 0.06792259 0.06792259
#>  [2,] 1.81624421 1.81624421 1.81624421
#>  [3,] 0.49069155 0.49069155 0.49069155
#>  [4,] 0.03826384 0.03826384 0.03826384
#>  [5,] 1.77949321 1.77949321 1.77949321
#>  [6,] 1.79116723 1.79116723 1.79116723
#>  [7,] 2.79436029 2.79436029 2.79436029
#>  [8,] 1.36500753 1.36500753 1.36500753
#>  [9,] 0.29860167 0.29860167 0.29860167
#> [10,] 0.87938504 0.87938504 0.87938504

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1412211 3.5591500 3.5591500
#>  [2,] 0.2887502 0.2887502 0.2887502
#>  [3,] 0.2639198 0.5283513 0.5283513
#>  [4,] 1.4505810 0.6891556 1.5774272
#>  [5,] 0.5605532 0.5605532 0.5605532
#>  [6,] 0.1864632 1.0406800 0.1864632
#>  [7,] 0.7462666 2.1009777 2.6126331
#>  [8,] 3.2191318 3.2191318 0.9130717
#>  [9,] 0.1620442 0.1620442 0.1620442
#> [10,] 1.4120278 1.0058423 0.5919195
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>             [,1]      [,2]         [,3]
#>  [1,] 0.31803892 0.6347741 1.6841762493
#>  [2,] 0.06943427 1.7858143 0.6555845961
#>  [3,] 0.35050161 1.4279439 0.6904764818
#>  [4,] 0.57509910 0.5438288 0.7620872669
#>  [5,] 1.04766122 0.1225577 0.0007156532
#>  [6,] 0.56551627 0.5576519 1.5047466644
#>  [7,] 2.84703387 0.5380492 0.8314765133
#>  [8,] 0.98037595 2.2903607 1.9269601516
#>  [9,] 1.45841499 1.9376603 0.3652534690
#> [10,] 1.45880989 1.5059568 0.7288836827
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.11216972 1.11216972 1.11216972
#>  [2,] 0.20296568 0.20296568 0.20296568
#>  [3,] 1.87661943 1.87661943 1.87661943
#>  [4,] 0.68930115 0.68930115 0.68930115
#>  [5,] 0.80172633 0.80172633 0.80172633
#>  [6,] 1.86754616 1.86754616 1.86754616
#>  [7,] 1.50408698 1.50408698 1.50408698
#>  [8,] 0.08208658 0.08208658 0.08208658
#>  [9,] 0.20209959 0.20209959 0.20209959
#> [10,] 4.04782886 4.04782886 4.04782886