Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.89024764 1.8485167 1.52415146
#>  [2,] 0.20679441 0.2067944 0.10416495
#>  [3,] 0.34416573 0.1373179 1.65197058
#>  [4,] 0.09626265 0.1145627 1.25163958
#>  [5,] 1.24608018 1.2460802 2.09428420
#>  [6,] 0.05934001 0.1910253 0.02036558
#>  [7,] 1.11365106 1.1136511 1.11365106
#>  [8,] 2.05440533 0.1758641 0.01509545
#>  [9,] 0.48279869 0.5569243 0.48279869
#> [10,] 1.37900372 2.0248860 1.37900372
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]      [,2]      [,3]
#>  [1,] 2.38946629 1.2208449 1.6986285
#>  [2,] 0.73640210 0.5205052 1.6380911
#>  [3,] 0.16099516 0.5749986 1.4228450
#>  [4,] 0.68014518 1.1716865 0.5863232
#>  [5,] 0.08565755 1.6673443 1.3053559
#>  [6,] 0.93246178 1.6652323 2.1434105
#>  [7,] 1.52423092 2.6216699 0.1665508
#>  [8,] 0.39052877 0.3870677 1.7009748
#>  [9,] 0.16585156 0.3509272 0.6822531
#> [10,] 6.64769521 0.8087286 1.3998218
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.80530777 1.80530777 1.80530777
#>  [2,] 1.45869138 1.45869138 1.45869138
#>  [3,] 0.71774235 0.71774235 0.71774235
#>  [4,] 0.97548954 0.97548954 0.97548954
#>  [5,] 2.51658875 2.51658875 2.51658875
#>  [6,] 2.32102972 2.32102972 2.32102972
#>  [7,] 0.13243079 0.13243079 0.13243079
#>  [8,] 0.03768946 0.03768946 0.03768946
#>  [9,] 2.27825866 2.27825866 2.27825866
#> [10,] 0.05663156 0.05663156 0.05663156

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3476218 1.7017465 2.02578444
#>  [2,] 0.7680209 0.7680209 0.76802089
#>  [3,] 0.5495040 0.7110127 0.71101272
#>  [4,] 0.6407783 0.5765063 0.57650629
#>  [5,] 0.3686398 0.5072858 0.50728579
#>  [6,] 2.0152375 0.0241343 0.00956941
#>  [7,] 1.7636619 0.4198135 1.76366191
#>  [8,] 0.5578830 2.5087993 0.88105748
#>  [9,] 0.5522995 0.8653226 0.86532256
#> [10,] 1.2070199 1.2070199 1.20701986
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]      [,2]       [,3]
#>  [1,] 2.84371652 0.7965503 0.27853339
#>  [2,] 0.88247760 0.2303477 0.98529022
#>  [3,] 3.82214891 1.1269056 4.20960994
#>  [4,] 0.37820605 1.1400108 0.14124127
#>  [5,] 4.57367263 1.5162093 0.13667947
#>  [6,] 0.32863056 0.4539031 0.46325826
#>  [7,] 1.98400205 0.7818490 0.47275334
#>  [8,] 0.72925168 0.8067252 0.04275028
#>  [9,] 0.04689777 0.9555441 0.98562367
#> [10,] 0.31691542 0.4701024 0.43825013
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6422655 1.6422655 1.6422655
#>  [2,] 3.7121398 3.7121398 3.7121398
#>  [3,] 0.8838151 0.8838151 0.8838151
#>  [4,] 0.5942577 0.5942577 0.5942577
#>  [5,] 1.0041369 1.0041369 1.0041369
#>  [6,] 1.0098766 1.0098766 1.0098766
#>  [7,] 0.4961176 0.4961176 0.4961176
#>  [8,] 2.3436655 2.3436655 2.3436655
#>  [9,] 0.5802323 0.5802323 0.5802323
#> [10,] 4.2420366 4.2420366 4.2420366

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.84398507 2.26643496 2.95810529
#>  [2,] 1.36439868 0.80345548 0.89822362
#>  [3,] 2.18242418 1.71641145 0.72576553
#>  [4,] 2.95823937 2.75292639 2.14739473
#>  [5,] 0.44724954 1.99810580 1.99810580
#>  [6,] 1.68921804 1.14095588 1.43370931
#>  [7,] 0.47462333 1.36057866 1.85317412
#>  [8,] 0.94820347 0.23579804 0.14276163
#>  [9,] 0.05543086 0.05543086 0.05543086
#> [10,] 0.35327770 0.35327770 5.15368088
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>              [,1]       [,2]       [,3]
#>  [1,] 1.163768111 0.50022734 1.41678940
#>  [2,] 0.233341120 0.04257452 0.08326916
#>  [3,] 0.632936747 0.19383787 1.36398803
#>  [4,] 1.862102488 0.08184880 1.68288375
#>  [5,] 1.210899027 0.03835468 1.04424795
#>  [6,] 0.925473745 3.00277619 1.29108493
#>  [7,] 0.139664546 0.06942562 0.55734279
#>  [8,] 0.574254776 0.33120906 1.21347951
#>  [9,] 0.491301293 0.73144219 0.49786475
#> [10,] 0.006273021 0.31883497 0.36259734
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02753466 0.02753466 0.02753466
#>  [2,] 1.78029703 1.78029703 1.78029703
#>  [3,] 0.98065070 0.98065070 0.98065070
#>  [4,] 0.05410342 0.05410342 0.05410342
#>  [5,] 3.04342833 3.04342833 3.04342833
#>  [6,] 0.66028497 0.66028497 0.66028497
#>  [7,] 1.61418486 1.61418486 1.61418486
#>  [8,] 1.01155602 1.01155602 1.01155602
#>  [9,] 1.15847963 1.15847963 1.15847963
#> [10,] 0.63715407 0.63715407 0.63715407