Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3076804 0.8449671 0.3076804
#>  [2,] 0.1301615 1.1432879 0.3009554
#>  [3,] 1.2322233 0.3383117 1.2322233
#>  [4,] 0.5660459 4.6201772 0.5660459
#>  [5,] 0.1222400 0.3680278 0.3680278
#>  [6,] 0.3240608 0.3240608 0.3240608
#>  [7,] 6.0071850 0.8785389 0.8999929
#>  [8,] 2.3982866 1.0584857 5.7364028
#>  [9,] 0.4760707 0.3048484 0.5038945
#> [10,] 1.8301498 2.2466639 1.0925082
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>              [,1]       [,2]       [,3]
#>  [1,] 0.571353199 0.22822016 0.21333186
#>  [2,] 0.243160975 1.02935549 0.08698808
#>  [3,] 1.113051972 0.46911625 1.40622842
#>  [4,] 0.597664548 0.08816256 2.89657781
#>  [5,] 1.760195711 0.75325024 0.08471292
#>  [6,] 1.176382622 0.10134401 0.95071307
#>  [7,] 1.651314524 1.06640318 1.67837599
#>  [8,] 0.001595636 1.61294877 0.80604617
#>  [9,] 3.731738844 2.23369599 2.02431078
#> [10,] 0.236597086 0.38704262 1.27455562
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.81978887 1.81978887 1.81978887
#>  [2,] 0.02317566 0.02317566 0.02317566
#>  [3,] 0.40531973 0.40531973 0.40531973
#>  [4,] 0.04970929 0.04970929 0.04970929
#>  [5,] 0.40413870 0.40413870 0.40413870
#>  [6,] 2.19283928 2.19283928 2.19283928
#>  [7,] 1.40854422 1.40854422 1.40854422
#>  [8,] 5.69819902 5.69819902 5.69819902
#>  [9,] 0.38956362 0.38956362 0.38956362
#> [10,] 0.66270930 0.66270930 0.66270930

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.7648100 0.5430754 0.04019133
#>  [2,] 1.5940690 1.6740543 0.88499344
#>  [3,] 2.7390654 2.7998457 0.16422216
#>  [4,] 0.7804581 1.3474194 0.07865634
#>  [5,] 0.2524093 0.2524093 0.25240932
#>  [6,] 0.0159460 1.0213721 2.46678848
#>  [7,] 1.0511151 1.0511151 1.05111511
#>  [8,] 0.3883982 1.0414045 1.16508890
#>  [9,] 0.6300613 0.6300613 0.63006129
#> [10,] 1.3995094 1.3995094 1.78211365
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3924275 0.66293973 0.67372949
#>  [2,] 0.7208021 0.64329947 1.18138265
#>  [3,] 1.3125443 0.71493333 0.07612496
#>  [4,] 0.1271552 3.13356045 0.89145939
#>  [5,] 1.3361304 0.64404381 0.82739667
#>  [6,] 0.4200662 0.36682758 0.41872079
#>  [7,] 0.3216290 0.03910073 2.88890781
#>  [8,] 0.4736685 0.64743261 0.11913883
#>  [9,] 2.6832802 0.23647832 0.19776847
#> [10,] 2.1165655 0.69180722 0.09553340
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07482593 0.07482593 0.07482593
#>  [2,] 0.13855086 0.13855086 0.13855086
#>  [3,] 0.60135186 0.60135186 0.60135186
#>  [4,] 0.31831457 0.31831457 0.31831457
#>  [5,] 1.77004666 1.77004666 1.77004666
#>  [6,] 0.98754431 0.98754431 0.98754431
#>  [7,] 2.57324942 2.57324942 2.57324942
#>  [8,] 0.58348594 0.58348594 0.58348594
#>  [9,] 0.20721003 0.20721003 0.20721003
#> [10,] 0.69695400 0.69695400 0.69695400

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.48668143 1.49729875 1.1293175
#>  [2,] 2.70554980 1.90045814 0.3086491
#>  [3,] 0.17279232 2.42140945 0.8317984
#>  [4,] 0.45453013 0.45453013 0.4545301
#>  [5,] 0.09705907 0.09705907 0.8800690
#>  [6,] 0.94611506 0.09291022 2.3566494
#>  [7,] 0.85180989 0.85180989 0.7190964
#>  [8,] 0.32570710 1.87942381 2.4880958
#>  [9,] 1.19310678 0.63306882 0.6330688
#> [10,] 0.01389583 0.58774459 0.3214484
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.20437352 1.1602719 2.9584392
#>  [2,] 2.48715048 0.3593074 0.5317109
#>  [3,] 0.17352384 0.2028186 0.4990862
#>  [4,] 1.32072970 2.7538035 0.4514996
#>  [5,] 1.40147255 0.0751363 3.2964934
#>  [6,] 0.66190358 2.7092092 0.3664977
#>  [7,] 0.20778426 0.3753831 0.1832783
#>  [8,] 0.05623813 0.1276822 2.8094895
#>  [9,] 0.63066527 0.7595366 0.2669592
#> [10,] 2.65111790 0.3013871 0.1640928
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1961440 0.1961440 0.1961440
#>  [2,] 0.5894823 0.5894823 0.5894823
#>  [3,] 2.0372251 2.0372251 2.0372251
#>  [4,] 0.3479716 0.3479716 0.3479716
#>  [5,] 0.6109167 0.6109167 0.6109167
#>  [6,] 2.5460539 2.5460539 2.5460539
#>  [7,] 0.9454817 0.9454817 0.9454817
#>  [8,] 0.4377284 0.4377284 0.4377284
#>  [9,] 1.0273561 1.0273561 1.0273561
#> [10,] 1.6122170 1.6122170 1.6122170