Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2163851 0.4531987 0.4531987
#>  [2,] 0.3471771 0.1885089 0.5053691
#>  [3,] 1.0837514 0.1473673 1.0837514
#>  [4,] 2.1131480 1.9544781 7.6302250
#>  [5,] 0.1700299 0.1700299 0.9919690
#>  [6,] 0.6862191 2.1021530 0.6862191
#>  [7,] 1.3477143 3.3105760 0.5851329
#>  [8,] 0.7305729 0.7305729 0.7305729
#>  [9,] 0.9216035 0.1134046 0.9216035
#> [10,] 0.9610444 1.1773493 1.1773493
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.20885197 0.7842507 1.20615073
#>  [2,] 1.46777125 1.1122058 0.84384571
#>  [3,] 1.87913973 1.1023833 1.32257683
#>  [4,] 0.78971986 0.2004759 2.50813768
#>  [5,] 1.11770988 0.2569275 1.48510343
#>  [6,] 1.38654237 4.5729333 0.58357212
#>  [7,] 0.21506664 0.1973649 1.40910783
#>  [8,] 0.41920724 1.9265055 1.07080772
#>  [9,] 0.85661173 0.3781733 0.16680108
#> [10,] 0.09583265 1.9148285 0.06557056
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02034434 0.02034434 0.02034434
#>  [2,] 0.86034568 0.86034568 0.86034568
#>  [3,] 4.38173710 4.38173710 4.38173710
#>  [4,] 0.46928754 0.46928754 0.46928754
#>  [5,] 0.30245824 0.30245824 0.30245824
#>  [6,] 0.32934946 0.32934946 0.32934946
#>  [7,] 0.56654207 0.56654207 0.56654207
#>  [8,] 0.28630987 0.28630987 0.28630987
#>  [9,] 0.08542691 0.08542691 0.08542691
#> [10,] 0.01434353 0.01434353 0.01434353

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 4.2173843 3.5364231 1.2570483
#>  [2,] 0.8148305 2.0089631 0.8949029
#>  [3,] 0.1192381 0.1192381 0.1192381
#>  [4,] 1.3098388 1.3098388 1.3813874
#>  [5,] 1.5121293 0.3622316 0.8358576
#>  [6,] 2.2990122 1.6962029 2.2990122
#>  [7,] 1.8690645 1.8690645 1.8690645
#>  [8,] 3.6261452 3.6261452 1.2055796
#>  [9,] 0.9001489 0.3484242 0.9001489
#> [10,] 0.9949739 0.3957322 2.4188421
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.07678425 0.6332649 0.12251687
#>  [2,] 0.63886101 0.4747842 1.67818420
#>  [3,] 1.03308064 0.4535207 0.86539846
#>  [4,] 0.23966501 0.7244258 0.77126553
#>  [5,] 1.73734171 0.5911769 1.87858410
#>  [6,] 1.05793344 0.6390671 0.17907150
#>  [7,] 2.06484389 1.3846317 0.11850615
#>  [8,] 1.63816735 1.4948761 0.07582565
#>  [9,] 2.18337249 0.2952324 0.74125360
#> [10,] 1.46744354 1.9322428 0.16255734
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.70214275 0.70214275 0.70214275
#>  [2,] 2.81687025 2.81687025 2.81687025
#>  [3,] 1.03437751 1.03437751 1.03437751
#>  [4,] 0.29254044 0.29254044 0.29254044
#>  [5,] 0.02228213 0.02228213 0.02228213
#>  [6,] 1.73744433 1.73744433 1.73744433
#>  [7,] 2.10109678 2.10109678 2.10109678
#>  [8,] 0.06879026 0.06879026 0.06879026
#>  [9,] 0.30280236 0.30280236 0.30280236
#> [10,] 0.01208125 0.01208125 0.01208125

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.25584898 2.17597513 3.10947098
#>  [2,] 0.45396557 0.48215760 0.42465236
#>  [3,] 0.09933984 0.09933984 0.09933984
#>  [4,] 0.43134731 0.47948163 0.47948163
#>  [5,] 0.22023456 0.22023456 0.22023456
#>  [6,] 0.67179370 1.03755792 0.23027175
#>  [7,] 0.34460452 0.12335688 0.34460452
#>  [8,] 0.48739573 1.60429518 0.78312637
#>  [9,] 1.48941493 1.48941493 2.21502180
#> [10,] 0.13035446 0.13035446 0.13035446
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24053604 1.42158745 0.01785103
#>  [2,] 1.89955826 0.25025392 1.86203927
#>  [3,] 0.50167782 1.20001869 3.19393889
#>  [4,] 0.14075723 0.96077715 0.66847258
#>  [5,] 1.16075491 0.57507057 2.14538030
#>  [6,] 0.04784156 1.15270123 0.01831163
#>  [7,] 0.17017187 0.06438595 0.03426558
#>  [8,] 0.27575175 2.62649569 0.53568768
#>  [9,] 2.12952669 1.41283615 0.67425002
#> [10,] 0.36348891 1.67659526 4.26241751
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46869492 0.46869492 0.46869492
#>  [2,] 0.73990653 0.73990653 0.73990653
#>  [3,] 0.20391413 0.20391413 0.20391413
#>  [4,] 0.05237211 0.05237211 0.05237211
#>  [5,] 1.24343939 1.24343939 1.24343939
#>  [6,] 0.24600229 0.24600229 0.24600229
#>  [7,] 0.35259237 0.35259237 0.35259237
#>  [8,] 0.23616599 0.23616599 0.23616599
#>  [9,] 0.92399612 0.92399612 0.92399612
#> [10,] 2.80460061 2.80460061 2.80460061