Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.10146576 0.10146576 0.1014658
#>  [2,] 0.26047718 0.02528749 0.2604772
#>  [3,] 2.95205568 2.95205568 1.0073156
#>  [4,] 3.04844591 1.59760973 2.9718476
#>  [5,] 0.51287832 0.12656364 0.1391212
#>  [6,] 2.39472188 1.63596406 2.3947219
#>  [7,] 0.03194953 4.57305048 2.3941488
#>  [8,] 0.90697155 0.90697155 0.9069715
#>  [9,] 1.90230183 1.90230183 1.6857525
#> [10,] 0.14092585 2.01718883 0.8791822
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>              [,1]       [,2]       [,3]
#>  [1,] 0.009380986 0.55034659 1.99036281
#>  [2,] 0.320233624 1.41892670 0.01538702
#>  [3,] 0.615567029 0.02120502 2.52527390
#>  [4,] 0.130624978 1.13559870 0.22679555
#>  [5,] 0.332503349 0.07884173 0.20910452
#>  [6,] 0.071651209 2.41185147 0.60716311
#>  [7,] 1.178532108 0.45653411 0.83005820
#>  [8,] 3.325885284 0.73933277 3.25663728
#>  [9,] 0.240153700 0.15013724 0.20945634
#> [10,] 1.391926324 0.57570793 0.24246981
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.14620239 2.14620239 2.14620239
#>  [2,] 0.07020489 0.07020489 0.07020489
#>  [3,] 3.05454640 3.05454640 3.05454640
#>  [4,] 1.42894355 1.42894355 1.42894355
#>  [5,] 0.91392534 0.91392534 0.91392534
#>  [6,] 1.39777804 1.39777804 1.39777804
#>  [7,] 0.86471119 0.86471119 0.86471119
#>  [8,] 0.68135837 0.68135837 0.68135837
#>  [9,] 0.60497138 0.60497138 0.60497138
#> [10,] 0.03497837 0.03497837 0.03497837

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.30699978 1.60073055 0.82324903
#>  [2,] 3.19338475 0.75637119 0.75637119
#>  [3,] 0.24284915 0.24284915 0.24284915
#>  [4,] 3.81939986 0.59921310 0.83711498
#>  [5,] 0.74852226 0.02474983 0.02474983
#>  [6,] 0.18029479 0.18029479 0.18029479
#>  [7,] 1.06129259 1.06129259 0.76181590
#>  [8,] 1.75022305 0.72001307 0.36196077
#>  [9,] 0.05859216 0.05859216 0.05859216
#> [10,] 0.16956129 0.69457098 0.77625876
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4191458 1.7854848 0.1169265
#>  [2,] 0.3330306 0.5301728 2.1872384
#>  [3,] 3.7900952 0.2921342 0.2901314
#>  [4,] 2.0449916 0.1495107 0.1973180
#>  [5,] 0.2085558 1.1313058 0.3502464
#>  [6,] 0.7883987 1.8064230 0.1389159
#>  [7,] 0.5807523 0.4778168 1.4968893
#>  [8,] 0.7369036 2.8219013 2.0432620
#>  [9,] 2.1874006 0.4187949 0.3543326
#> [10,] 0.5149973 1.5565283 1.4554846
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9943392 0.9943392 0.9943392
#>  [2,] 0.7864131 0.7864131 0.7864131
#>  [3,] 0.6918684 0.6918684 0.6918684
#>  [4,] 0.8557860 0.8557860 0.8557860
#>  [5,] 1.3717457 1.3717457 1.3717457
#>  [6,] 0.5102923 0.5102923 0.5102923
#>  [7,] 1.0169345 1.0169345 1.0169345
#>  [8,] 3.0794179 3.0794179 3.0794179
#>  [9,] 1.7375999 1.7375999 1.7375999
#> [10,] 1.3457851 1.3457851 1.3457851

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.02057023 0.020570232 0.02057023
#>  [2,] 0.39756415 0.125155511 0.12515551
#>  [3,] 2.47538971 2.113890818 0.05149121
#>  [4,] 1.02201947 1.022019472 1.02201947
#>  [5,] 0.34498862 0.344988617 0.34498862
#>  [6,] 0.22407097 0.196571473 0.19657147
#>  [7,] 0.66390717 0.663907171 0.66390717
#>  [8,] 0.54267874 0.186781920 0.94225113
#>  [9,] 0.53119080 0.531190804 0.53119080
#> [10,] 0.12930217 0.009634955 0.12930217
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]      [,2]       [,3]
#>  [1,] 2.09343960 0.5472794 2.74064491
#>  [2,] 0.39617214 0.9919160 0.04598757
#>  [3,] 1.63399440 0.9247774 0.10791145
#>  [4,] 0.46413728 0.1333974 1.10270805
#>  [5,] 1.57500216 0.1629161 0.45180478
#>  [6,] 0.49696649 0.7945794 0.13731468
#>  [7,] 0.65183373 0.2853772 1.23107997
#>  [8,] 1.03323490 0.6934460 1.95285700
#>  [9,] 1.69952146 0.4076147 0.63883288
#> [10,] 0.05813571 0.2502493 2.11755502
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.72210901 3.72210901 3.72210901
#>  [2,] 0.64269537 0.64269537 0.64269537
#>  [3,] 0.29745821 0.29745821 0.29745821
#>  [4,] 0.13806670 0.13806670 0.13806670
#>  [5,] 2.66083744 2.66083744 2.66083744
#>  [6,] 0.01869945 0.01869945 0.01869945
#>  [7,] 0.31088567 0.31088567 0.31088567
#>  [8,] 1.15198766 1.15198766 1.15198766
#>  [9,] 1.10466379 1.10466379 1.10466379
#> [10,] 0.14060957 0.14060957 0.14060957