Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.68198671 0.68198671 2.45586760
#>  [2,] 1.34561286 1.43296891 1.34561286
#>  [3,] 1.25894161 1.25894161 0.12767775
#>  [4,] 0.45799633 0.45799633 0.45799633
#>  [5,] 0.26680313 1.32668597 1.32668597
#>  [6,] 5.71087509 1.07617313 0.59030150
#>  [7,] 0.99671492 1.42434905 0.80287811
#>  [8,] 0.04094346 0.04094346 1.42723194
#>  [9,] 0.61002586 0.11244424 0.13331228
#> [10,] 0.99891274 0.45452910 0.08402567
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>            [,1]       [,2]       [,3]
#>  [1,] 2.8952670 0.10049523 3.06375075
#>  [2,] 0.3352016 1.47797247 1.67250631
#>  [3,] 0.9558737 1.07683543 0.80503151
#>  [4,] 0.1893903 2.31548352 0.74247069
#>  [5,] 0.2054264 0.36471227 0.61225156
#>  [6,] 0.1409252 0.05136667 0.15799080
#>  [7,] 0.4044552 3.09556062 0.01352914
#>  [8,] 0.2801275 0.99319912 4.04173282
#>  [9,] 0.1471760 0.99131101 0.25091132
#> [10,] 2.3250782 0.28759328 0.73222216
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03851786 0.03851786 0.03851786
#>  [2,] 0.12730557 0.12730557 0.12730557
#>  [3,] 0.68482832 0.68482832 0.68482832
#>  [4,] 0.06285168 0.06285168 0.06285168
#>  [5,] 0.58168486 0.58168486 0.58168486
#>  [6,] 0.66765243 0.66765243 0.66765243
#>  [7,] 0.76067101 0.76067101 0.76067101
#>  [8,] 0.07637940 0.07637940 0.07637940
#>  [9,] 0.86360386 0.86360386 0.86360386
#> [10,] 2.59858965 2.59858965 2.59858965

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2102824 0.6469235 0.64692351
#>  [2,] 4.0826704 1.7059437 4.08267036
#>  [3,] 1.8102573 1.4085128 1.73879576
#>  [4,] 0.7858875 0.1197051 0.78588754
#>  [5,] 0.2540050 0.2540050 0.25400505
#>  [6,] 0.3480906 2.7043449 0.02377851
#>  [7,] 0.4204877 0.4204877 0.42048768
#>  [8,] 1.5528816 0.3237609 0.95948141
#>  [9,] 0.4459958 0.7159891 0.71598909
#> [10,] 1.5002054 1.2430801 2.45165022
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.6823508 0.6654172 0.60564368
#>  [2,] 1.3234221 0.7545664 0.01221926
#>  [3,] 2.7034363 0.7159494 0.07685029
#>  [4,] 0.3413766 0.1929695 0.15354108
#>  [5,] 0.1994870 0.1372084 1.52809180
#>  [6,] 0.2671238 0.2837678 0.04914597
#>  [7,] 0.1411772 1.1497462 1.10178462
#>  [8,] 1.4993477 1.1751128 1.77186882
#>  [9,] 0.0808355 0.2337571 0.28160491
#> [10,] 1.4278111 0.2857455 0.10956072
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84751581 0.84751581 0.84751581
#>  [2,] 0.07823443 0.07823443 0.07823443
#>  [3,] 1.04936916 1.04936916 1.04936916
#>  [4,] 0.31583974 0.31583974 0.31583974
#>  [5,] 0.04088148 0.04088148 0.04088148
#>  [6,] 0.26974771 0.26974771 0.26974771
#>  [7,] 0.69891945 0.69891945 0.69891945
#>  [8,] 0.03310455 0.03310455 0.03310455
#>  [9,] 0.72260674 0.72260674 0.72260674
#> [10,] 0.27217529 0.27217529 0.27217529

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.836354345 2.0540819 2.8447924
#>  [2,] 3.569862735 1.7043941 1.2440143
#>  [3,] 0.067116602 0.0671166 0.1153366
#>  [4,] 0.773269369 1.6946849 2.0527960
#>  [5,] 0.286897857 0.2868979 0.2868979
#>  [6,] 0.938084012 0.1983848 0.1983848
#>  [7,] 0.106864920 0.1068649 0.1068649
#>  [8,] 1.175007465 0.1868874 1.1750075
#>  [9,] 0.002500347 0.1736597 0.1736597
#> [10,] 1.523131344 1.5231313 1.0653152
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5143820 0.2508439 1.3314538
#>  [2,] 0.7853668 1.4761040 0.6019317
#>  [3,] 1.2442720 3.0786133 0.8096371
#>  [4,] 0.3541331 1.5468309 0.2369371
#>  [5,] 0.1262416 1.0107728 0.8736933
#>  [6,] 0.2774701 0.1553745 0.8908122
#>  [7,] 0.4214570 1.4424291 0.6532703
#>  [8,] 0.3643962 0.1641358 1.3838830
#>  [9,] 0.2477984 3.0825447 0.6700498
#> [10,] 1.2621672 0.4564108 0.5359938
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.83856938 1.83856938 1.83856938
#>  [2,] 1.16067360 1.16067360 1.16067360
#>  [3,] 3.24921353 3.24921353 3.24921353
#>  [4,] 0.18786590 0.18786590 0.18786590
#>  [5,] 0.19356180 0.19356180 0.19356180
#>  [6,] 0.57276779 0.57276779 0.57276779
#>  [7,] 0.61074993 0.61074993 0.61074993
#>  [8,] 2.77800226 2.77800226 2.77800226
#>  [9,] 0.05007757 0.05007757 0.05007757
#> [10,] 0.75995518 0.75995518 0.75995518