Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3148191 0.12278150 3.0281170
#>  [2,] 0.5520722 0.30817371 1.8923781
#>  [3,] 0.3103666 0.31036657 0.3103666
#>  [4,] 2.2887231 0.04666468 1.9206816
#>  [5,] 1.6212589 0.52313448 0.5231345
#>  [6,] 1.2031833 0.80875242 1.9412192
#>  [7,] 1.3505046 0.78468423 1.4051691
#>  [8,] 1.8412439 0.67293773 0.6729377
#>  [9,] 0.7221265 0.72212652 0.7221265
#> [10,] 0.1854833 0.18548331 0.1854833
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6405765 0.16335754 0.22351113
#>  [2,] 0.8934731 0.76082096 0.85322151
#>  [3,] 0.5130370 1.23370376 0.12781737
#>  [4,] 0.3330734 0.60803535 1.84283388
#>  [5,] 2.4754276 1.42149911 0.12207202
#>  [6,] 0.9011766 0.98673288 1.15809776
#>  [7,] 0.2393643 0.08749957 0.02897022
#>  [8,] 0.9527948 2.20393800 0.56533745
#>  [9,] 0.0796650 0.06401505 0.24378770
#> [10,] 0.3536203 0.25594359 2.44764759
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9562064 0.9562064 0.9562064
#>  [2,] 0.4257754 0.4257754 0.4257754
#>  [3,] 3.2947522 3.2947522 3.2947522
#>  [4,] 1.0631760 1.0631760 1.0631760
#>  [5,] 1.4348807 1.4348807 1.4348807
#>  [6,] 0.0630729 0.0630729 0.0630729
#>  [7,] 0.1983240 0.1983240 0.1983240
#>  [8,] 2.1777504 2.1777504 2.1777504
#>  [9,] 0.1916578 0.1916578 0.1916578
#> [10,] 1.2508610 1.2508610 1.2508610

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.31787870 1.94327665 0.3178787
#>  [2,] 0.57528176 2.17817060 2.1781706
#>  [3,] 0.07505619 2.30130479 0.6558727
#>  [4,] 0.42938858 0.42938858 1.0191441
#>  [5,] 0.37370734 0.37370734 1.2232524
#>  [6,] 0.64865596 0.09208419 0.6486560
#>  [7,] 1.64511889 0.21648848 0.5880591
#>  [8,] 0.37979845 0.37979845 0.3797985
#>  [9,] 3.81751308 1.41083613 0.8805657
#> [10,] 0.81523658 1.30673214 0.8152366
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.85095565 1.89496061 1.68737598
#>  [2,] 0.29530690 0.40537393 0.91071715
#>  [3,] 2.20085349 0.03626795 0.91932739
#>  [4,] 2.10341867 0.35953902 0.69763770
#>  [5,] 0.55872303 0.31005409 1.01470461
#>  [6,] 0.36045693 0.34170542 0.06189610
#>  [7,] 1.74625869 1.43087828 0.37507316
#>  [8,] 0.07063891 0.46520883 0.33364764
#>  [9,] 0.02872629 0.33057286 0.06151487
#> [10,] 0.08458493 0.13991322 1.46518798
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.28489651 1.28489651 1.28489651
#>  [2,] 0.08085339 0.08085339 0.08085339
#>  [3,] 2.57689410 2.57689410 2.57689410
#>  [4,] 2.98976837 2.98976837 2.98976837
#>  [5,] 2.99475648 2.99475648 2.99475648
#>  [6,] 0.77259632 0.77259632 0.77259632
#>  [7,] 1.88952182 1.88952182 1.88952182
#>  [8,] 7.82619385 7.82619385 7.82619385
#>  [9,] 1.44459334 1.44459334 1.44459334
#> [10,] 0.48375155 0.48375155 0.48375155

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>               [,1]       [,2]      [,3]
#>  [1,] 1.7822059937 3.52696886 3.0126353
#>  [2,] 0.8297308804 0.96144029 0.3216448
#>  [3,] 0.0009654387 0.25345937 0.2534594
#>  [4,] 0.6165360076 0.53556653 0.6165360
#>  [5,] 0.7408626097 0.45135895 0.7408626
#>  [6,] 0.8007969607 0.36080831 0.8007970
#>  [7,] 1.3633546555 0.55751214 1.3633547
#>  [8,] 1.1054167666 0.69827208 1.2664015
#>  [9,] 0.9304057813 1.11428333 3.0783949
#> [10,] 3.4217520289 0.06822372 1.6049057
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.66883327 1.72033286 0.1605650
#>  [2,] 1.33580511 0.06887378 0.2908567
#>  [3,] 0.64383847 0.11847223 1.7637633
#>  [4,] 1.05347545 0.00974876 0.1644333
#>  [5,] 2.82769863 2.59556250 0.6510010
#>  [6,] 0.04518029 0.32460170 0.2269538
#>  [7,] 1.06309959 0.68994061 0.1200139
#>  [8,] 2.74867608 0.73897802 1.9523734
#>  [9,] 1.66407918 1.45252369 0.2377435
#> [10,] 0.22276772 1.27124617 1.1187123
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2181881 0.2181881 0.2181881
#>  [2,] 3.6769128 3.6769128 3.6769128
#>  [3,] 1.7690621 1.7690621 1.7690621
#>  [4,] 0.1337285 0.1337285 0.1337285
#>  [5,] 0.9089768 0.9089768 0.9089768
#>  [6,] 0.6186524 0.6186524 0.6186524
#>  [7,] 1.1584534 1.1584534 1.1584534
#>  [8,] 1.2558608 1.2558608 1.2558608
#>  [9,] 0.6258146 0.6258146 0.6258146
#> [10,] 0.4082268 0.4082268 0.4082268