Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23215947 1.69467794 0.23215947
#>  [2,] 2.92370614 1.93629959 3.05427003
#>  [3,] 0.02637644 0.02637644 0.02637644
#>  [4,] 2.63744848 0.83984085 2.00700183
#>  [5,] 1.52090545 1.37339376 1.52090545
#>  [6,] 0.78500572 0.78500572 0.78500572
#>  [7,] 0.40032527 0.77369547 1.57554587
#>  [8,] 0.08185970 0.78189355 0.49687875
#>  [9,] 0.77484158 2.67047086 0.06273591
#> [10,] 0.75504338 0.75504338 0.75504338
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.68300917 0.04556962 1.1691946
#>  [2,] 2.45212506 0.34575415 1.7669136
#>  [3,] 0.74971354 1.29725754 2.2339877
#>  [4,] 0.30945986 0.09440005 0.1884293
#>  [5,] 0.04946133 0.59237587 0.7496497
#>  [6,] 3.23788384 1.23877284 0.7400603
#>  [7,] 0.03181765 0.19325333 0.5447401
#>  [8,] 0.46295771 3.88471876 1.2704680
#>  [9,] 0.46371829 1.44797620 0.3316552
#> [10,] 0.30472910 0.00593829 0.6543347
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.28794829 1.28794829 1.28794829
#>  [2,] 0.82827013 0.82827013 0.82827013
#>  [3,] 0.08802441 0.08802441 0.08802441
#>  [4,] 1.42349738 1.42349738 1.42349738
#>  [5,] 0.54151460 0.54151460 0.54151460
#>  [6,] 0.60213191 0.60213191 0.60213191
#>  [7,] 1.70192057 1.70192057 1.70192057
#>  [8,] 0.02035929 0.02035929 0.02035929
#>  [9,] 0.75933730 0.75933730 0.75933730
#> [10,] 1.83955425 1.83955425 1.83955425

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21043625 0.21043625 0.21043625
#>  [2,] 1.21841674 1.21841674 1.21841674
#>  [3,] 0.44322906 1.69109587 1.69109587
#>  [4,] 0.90329626 2.47697369 2.47697369
#>  [5,] 3.75016242 3.45142902 5.03808701
#>  [6,] 0.18366735 0.02089972 0.02089972
#>  [7,] 1.77570567 0.20132532 2.58260040
#>  [8,] 0.09728449 0.09728449 0.09728449
#>  [9,] 0.95165557 0.95165557 2.15430552
#> [10,] 1.79438819 1.79438819 1.79438819
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>            [,1]         [,2]      [,3]
#>  [1,] 0.7163837 0.8284923854 0.1828490
#>  [2,] 2.8610839 0.0004960517 0.6298298
#>  [3,] 1.0289251 0.7939973958 0.8028825
#>  [4,] 0.8488407 1.2500831678 0.2246282
#>  [5,] 0.6859288 0.0143762776 0.4770640
#>  [6,] 2.9823195 0.0011463007 0.5770682
#>  [7,] 1.8352273 1.9276085938 1.3190866
#>  [8,] 0.4259377 0.5782550389 0.4958176
#>  [9,] 1.3045980 2.7886855138 0.1708515
#> [10,] 0.1583211 3.0405044945 0.6110450
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9886019 0.9886019 0.9886019
#>  [2,] 4.6967083 4.6967083 4.6967083
#>  [3,] 0.1603633 0.1603633 0.1603633
#>  [4,] 0.1010703 0.1010703 0.1010703
#>  [5,] 0.3037485 0.3037485 0.3037485
#>  [6,] 1.4033419 1.4033419 1.4033419
#>  [7,] 0.7923410 0.7923410 0.7923410
#>  [8,] 0.8309203 0.8309203 0.8309203
#>  [9,] 0.6064395 0.6064395 0.6064395
#> [10,] 0.4487773 0.4487773 0.4487773

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08113458 0.08113458 0.08113458
#>  [2,] 0.97991472 0.97991472 1.06039440
#>  [3,] 2.65028773 3.17062432 0.67652024
#>  [4,] 0.29888304 0.29888304 0.29888304
#>  [5,] 0.77403609 3.48725507 0.76723400
#>  [6,] 1.42975675 1.92638829 0.10042808
#>  [7,] 0.36820345 0.75759726 0.91302929
#>  [8,] 0.28750280 0.28750280 0.28750280
#>  [9,] 0.29171100 1.00626153 0.28151125
#> [10,] 0.88692581 0.88692581 2.16448886
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]       [,2]         [,3]
#>  [1,] 1.0422063 1.82379671 0.8757011508
#>  [2,] 0.1584870 0.64007706 1.2324813790
#>  [3,] 0.9103868 0.50130577 0.0005143916
#>  [4,] 0.5965515 6.29922514 0.2062703585
#>  [5,] 1.4376837 2.56963340 1.5001581348
#>  [6,] 1.0161789 0.31902376 0.5212672027
#>  [7,] 0.6735015 0.14138459 0.7047381877
#>  [8,] 0.5900005 0.03035116 1.8364910848
#>  [9,] 0.4145360 0.78421230 0.4276689357
#> [10,] 2.0843297 1.07460562 0.7809164952
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4762074 0.4762074 0.4762074
#>  [2,] 1.2247087 1.2247087 1.2247087
#>  [3,] 0.1823321 0.1823321 0.1823321
#>  [4,] 0.6118016 0.6118016 0.6118016
#>  [5,] 1.1932805 1.1932805 1.1932805
#>  [6,] 0.2458695 0.2458695 0.2458695
#>  [7,] 0.1983748 0.1983748 0.1983748
#>  [8,] 0.5270769 0.5270769 0.5270769
#>  [9,] 0.4027729 0.4027729 0.4027729
#> [10,] 0.6870047 0.6870047 0.6870047