Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3025966 0.3025966 0.78244944
#>  [2,] 1.5583957 1.5583957 1.66697527
#>  [3,] 0.4792483 0.8403902 1.00730761
#>  [4,] 1.7891312 0.6041642 0.17757449
#>  [5,] 1.3911974 0.3776082 1.06834871
#>  [6,] 1.1899793 1.1899793 1.18997929
#>  [7,] 1.2133786 0.4837344 3.57006975
#>  [8,] 1.6944125 0.2854572 0.77001876
#>  [9,] 0.3357141 0.3357141 0.09827455
#> [10,] 1.7236452 1.7236452 0.01631878
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>             [,1]      [,2]      [,3]
#>  [1,] 0.18503518 0.1697226 0.7278774
#>  [2,] 0.63208513 0.4895517 0.3619899
#>  [3,] 0.79375759 0.4209031 1.4277919
#>  [4,] 0.55302101 1.6091147 0.5420319
#>  [5,] 0.97709704 0.7605021 0.1684545
#>  [6,] 0.02001029 1.4321040 0.2107280
#>  [7,] 0.93372540 0.6129883 0.7123773
#>  [8,] 0.95914726 2.3600465 0.1821822
#>  [9,] 0.82402450 1.5756797 0.6306210
#> [10,] 0.22360384 2.9365484 0.9966022
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.61110506 0.61110506 0.61110506
#>  [2,] 0.23463525 0.23463525 0.23463525
#>  [3,] 0.18631418 0.18631418 0.18631418
#>  [4,] 0.53903455 0.53903455 0.53903455
#>  [5,] 0.09737065 0.09737065 0.09737065
#>  [6,] 1.65756278 1.65756278 1.65756278
#>  [7,] 1.11030648 1.11030648 1.11030648
#>  [8,] 0.13683085 0.13683085 0.13683085
#>  [9,] 0.07168034 0.07168034 0.07168034
#> [10,] 0.99906544 0.99906544 0.99906544

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>            [,1]      [,2]         [,3]
#>  [1,] 0.2843856 0.9721787 1.0804350558
#>  [2,] 0.1102364 0.6513488 0.6513487576
#>  [3,] 0.3881419 0.3881419 0.0003307862
#>  [4,] 1.1494934 1.3887921 1.1494933942
#>  [5,] 0.2921265 0.2921265 0.4681532824
#>  [6,] 0.3678036 0.3678036 0.3678035857
#>  [7,] 0.8961770 0.8961770 0.2296760339
#>  [8,] 1.3972856 1.2604685 1.2080102380
#>  [9,] 2.2372384 2.2372384 1.1705060906
#> [10,] 1.3874891 0.4126546 0.4126546350
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.35074911 0.3386916 2.95162288
#>  [2,] 3.02359656 0.3742522 0.12568298
#>  [3,] 0.05458128 2.1298996 3.41064838
#>  [4,] 0.21518437 1.7555838 0.58547982
#>  [5,] 1.34094761 0.6801477 2.87941701
#>  [6,] 0.02968837 0.4571563 0.06256933
#>  [7,] 0.55012062 2.9694677 1.66987515
#>  [8,] 0.83052964 0.8656934 0.42374402
#>  [9,] 0.87906596 1.2789809 1.12346443
#> [10,] 0.90570938 0.5236253 0.26033571
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.78719021 0.78719021 0.78719021
#>  [2,] 0.89635491 0.89635491 0.89635491
#>  [3,] 0.85492006 0.85492006 0.85492006
#>  [4,] 0.05402046 0.05402046 0.05402046
#>  [5,] 0.46430558 0.46430558 0.46430558
#>  [6,] 2.15188215 2.15188215 2.15188215
#>  [7,] 0.21424607 0.21424607 0.21424607
#>  [8,] 0.78546194 0.78546194 0.78546194
#>  [9,] 1.70542881 1.70542881 1.70542881
#> [10,] 0.81317529 0.81317529 0.81317529

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6807726 0.6807726 0.68077256
#>  [2,] 0.2925728 0.2925728 0.29257279
#>  [3,] 0.1086233 0.6604895 0.10862329
#>  [4,] 2.1098599 1.3836600 2.53802423
#>  [5,] 0.6885609 0.6885609 0.48934589
#>  [6,] 0.8048108 0.8048108 0.05600244
#>  [7,] 0.1395907 0.1395907 0.13959074
#>  [8,] 1.4983360 1.4232417 0.80344292
#>  [9,] 0.9462328 0.9462328 0.94623283
#> [10,] 0.7411530 1.7949834 1.11269955
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1639969 0.07943750 3.0305827
#>  [2,] 0.3423018 2.53956485 2.2317994
#>  [3,] 0.1841056 0.18249737 0.3923088
#>  [4,] 1.9048412 1.14902556 0.4801023
#>  [5,] 1.9872007 0.84436196 2.6126096
#>  [6,] 2.6223397 0.92257250 1.0625832
#>  [7,] 0.3766490 0.77456883 0.3559819
#>  [8,] 0.1839364 0.06800547 0.5051772
#>  [9,] 0.9512950 0.70288595 1.4725121
#> [10,] 1.4512143 1.29410007 0.6415270
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04922431 0.04922431 0.04922431
#>  [2,] 0.05312340 0.05312340 0.05312340
#>  [3,] 1.59975268 1.59975268 1.59975268
#>  [4,] 2.33766256 2.33766256 2.33766256
#>  [5,] 1.19355965 1.19355965 1.19355965
#>  [6,] 1.23739168 1.23739168 1.23739168
#>  [7,] 0.74403686 0.74403686 0.74403686
#>  [8,] 1.21115909 1.21115909 1.21115909
#>  [9,] 0.42430595 0.42430595 0.42430595
#> [10,] 0.86748101 0.86748101 0.86748101