Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>            [,1]      [,2]       [,3]
#>  [1,] 2.2006843 2.2006843 0.09265645
#>  [2,] 0.3687940 0.4937123 0.36879401
#>  [3,] 1.7986336 1.2349211 0.05256958
#>  [4,] 0.7901493 2.1810884 1.39259700
#>  [5,] 3.0233746 1.7990468 1.62453014
#>  [6,] 4.2597790 0.5285621 0.52856207
#>  [7,] 2.8202982 0.3608124 1.01166750
#>  [8,] 0.5873549 0.5873549 0.46129619
#>  [9,] 2.2559338 2.3222195 2.25593383
#> [10,] 1.4160652 0.7470752 0.06993407
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.96238633 1.16402974 0.44008405
#>  [2,] 1.69963857 1.68838697 3.51431156
#>  [3,] 1.40391659 0.03138997 0.09824662
#>  [4,] 0.32148477 0.74004709 0.45378044
#>  [5,] 4.04812238 0.94747948 1.11758450
#>  [6,] 0.14600473 0.03775907 0.44832101
#>  [7,] 2.01627300 0.08363434 0.11164819
#>  [8,] 1.67636304 0.26044703 2.67903049
#>  [9,] 0.01819022 0.51201805 0.19530723
#> [10,] 0.18986250 1.33700805 0.06522249
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2906026 0.2906026 0.2906026
#>  [2,] 1.0703266 1.0703266 1.0703266
#>  [3,] 0.1052173 0.1052173 0.1052173
#>  [4,] 1.2021540 1.2021540 1.2021540
#>  [5,] 1.0461150 1.0461150 1.0461150
#>  [6,] 1.8697112 1.8697112 1.8697112
#>  [7,] 0.2077975 0.2077975 0.2077975
#>  [8,] 2.7424956 2.7424956 2.7424956
#>  [9,] 0.7179833 0.7179833 0.7179833
#> [10,] 1.5611554 1.5611554 1.5611554

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.01672897 0.18664079 0.38690048
#>  [2,] 0.68494241 2.11468803 0.78479118
#>  [3,] 1.06069116 0.34594908 1.06069116
#>  [4,] 0.69285973 1.43562791 1.37032663
#>  [5,] 0.50408426 0.50408426 0.50408426
#>  [6,] 0.24908498 0.24908498 0.79872861
#>  [7,] 1.27096872 0.07356381 2.91932836
#>  [8,] 0.03275745 1.53687759 1.53687759
#>  [9,] 1.10797424 2.67617764 0.08224562
#> [10,] 0.29518870 1.54986832 0.99509255
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.03508441 0.6006718 0.01508779
#>  [2,] 0.13299160 1.5692094 0.05478579
#>  [3,] 4.26653377 1.3390775 0.86391158
#>  [4,] 0.22258681 0.0524059 2.47900736
#>  [5,] 0.11185473 1.8473211 0.38333527
#>  [6,] 1.57934810 3.0342820 1.06716840
#>  [7,] 9.18700010 1.8479789 0.15489221
#>  [8,] 1.21395672 1.3538347 1.28306912
#>  [9,] 0.52977560 0.6209332 1.44138565
#> [10,] 0.87207956 1.3975999 0.12659489
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.18861230 2.18861230 2.18861230
#>  [2,] 1.14616015 1.14616015 1.14616015
#>  [3,] 2.14999835 2.14999835 2.14999835
#>  [4,] 0.25641479 0.25641479 0.25641479
#>  [5,] 0.15489578 0.15489578 0.15489578
#>  [6,] 0.13069032 0.13069032 0.13069032
#>  [7,] 4.99550860 4.99550860 4.99550860
#>  [8,] 0.06787707 0.06787707 0.06787707
#>  [9,] 0.36663991 0.36663991 0.36663991
#> [10,] 0.34796856 0.34796856 0.34796856

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0830426 0.84490506 3.02700608
#>  [2,] 0.3615337 0.02279786 0.36153366
#>  [3,] 0.8097678 2.81977081 0.53870823
#>  [4,] 3.0864168 0.27202242 0.22383941
#>  [5,] 1.0128429 1.01284286 1.01284286
#>  [6,] 3.0117448 0.00293198 1.69469716
#>  [7,] 3.0489987 2.02117621 2.02117621
#>  [8,] 0.8254594 0.82545940 0.82545940
#>  [9,] 0.8904027 1.14350001 0.25608185
#> [10,] 0.5040296 0.50402964 0.09552935
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 3.27753281 0.91372883 1.57313474
#>  [2,] 1.15443867 0.38669035 0.12961404
#>  [3,] 0.52276346 0.88722585 0.57105660
#>  [4,] 0.95803206 0.70147737 0.51102535
#>  [5,] 3.91970042 0.95676281 1.23552220
#>  [6,] 2.13047773 0.35961842 0.03291188
#>  [7,] 0.11126261 0.95257179 0.11010922
#>  [8,] 1.36130421 0.01278399 1.86935726
#>  [9,] 0.08909466 0.41418395 0.44309541
#> [10,] 0.03484187 1.47347899 0.63785243
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.94955782 2.94955782 2.94955782
#>  [2,] 2.21453176 2.21453176 2.21453176
#>  [3,] 0.40774810 0.40774810 0.40774810
#>  [4,] 0.40925453 0.40925453 0.40925453
#>  [5,] 0.23055791 0.23055791 0.23055791
#>  [6,] 0.98917721 0.98917721 0.98917721
#>  [7,] 0.03100206 0.03100206 0.03100206
#>  [8,] 0.88050654 0.88050654 0.88050654
#>  [9,] 2.80468930 2.80468930 2.80468930
#> [10,] 3.78233649 3.78233649 3.78233649