Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2964957 1.3076301 0.86940735
#>  [2,] 0.8499419 0.8499419 0.02500647
#>  [3,] 0.2269844 6.1620116 1.22971118
#>  [4,] 0.1598660 0.1598660 0.15986595
#>  [5,] 0.3046813 0.3046813 0.30468129
#>  [6,] 5.1503223 5.1503223 0.17271454
#>  [7,] 0.9638163 0.7712053 0.77120534
#>  [8,] 0.7322774 1.3732918 1.90662081
#>  [9,] 4.4044482 5.4409568 1.89880635
#> [10,] 1.1259377 0.9732337 0.97323370
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>             [,1]      [,2]       [,3]
#>  [1,] 1.20576309 0.0412045 0.40180252
#>  [2,] 1.30097299 0.6735468 0.94588372
#>  [3,] 3.16438210 3.5964853 0.23577113
#>  [4,] 1.59011579 3.1061684 1.04794451
#>  [5,] 1.85264461 2.1721513 0.18257312
#>  [6,] 2.27987247 0.7086900 0.64997107
#>  [7,] 0.01111566 0.6005539 0.28454875
#>  [8,] 0.65038485 1.3194884 0.07585563
#>  [9,] 0.27986259 0.2365574 2.67213641
#> [10,] 0.89706319 0.1187657 0.13477574
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3463818 0.3463818 0.3463818
#>  [2,] 0.2851037 0.2851037 0.2851037
#>  [3,] 1.6016998 1.6016998 1.6016998
#>  [4,] 0.1778639 0.1778639 0.1778639
#>  [5,] 2.0199385 2.0199385 2.0199385
#>  [6,] 0.2075630 0.2075630 0.2075630
#>  [7,] 0.1239507 0.1239507 0.1239507
#>  [8,] 1.2229626 1.2229626 1.2229626
#>  [9,] 1.2402341 1.2402341 1.2402341
#> [10,] 1.6350365 1.6350365 1.6350365

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.86890014 0.37415050 0.7867583
#>  [2,] 0.12241833 0.42788030 0.2301081
#>  [3,] 0.21734214 0.42858481 0.3915499
#>  [4,] 4.92970468 0.23495951 4.9297047
#>  [5,] 0.01658879 0.44673934 0.9579417
#>  [6,] 2.30295831 1.48701114 1.4870111
#>  [7,] 0.46858843 0.46858843 0.4685884
#>  [8,] 0.90508483 0.75892775 0.9050848
#>  [9,] 1.43988961 0.07902609 2.4955199
#> [10,] 1.09273993 0.08983818 1.3310508
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.05725591 0.046315210 0.54359734
#>  [2,] 2.44434051 0.441733259 0.59783219
#>  [3,] 0.38689746 0.498380997 1.00286988
#>  [4,] 1.04412348 1.456583190 0.32838062
#>  [5,] 1.44760131 1.147296679 1.21603056
#>  [6,] 2.11812039 0.583341073 0.71009514
#>  [7,] 0.08642943 0.106193961 1.28739397
#>  [8,] 0.77086661 6.007560887 1.59352549
#>  [9,] 1.63653736 0.937816479 0.82047689
#> [10,] 1.43948801 0.002878733 0.04907406
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1806599 0.1806599 0.1806599
#>  [2,] 4.7945331 4.7945331 4.7945331
#>  [3,] 0.6680044 0.6680044 0.6680044
#>  [4,] 1.1538952 1.1538952 1.1538952
#>  [5,] 0.5656287 0.5656287 0.5656287
#>  [6,] 0.0065551 0.0065551 0.0065551
#>  [7,] 1.2121218 1.2121218 1.2121218
#>  [8,] 0.2026583 0.2026583 0.2026583
#>  [9,] 0.2996026 0.2996026 0.2996026
#> [10,] 2.2478711 2.2478711 2.2478711

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.14672840 1.5657794 1.67018289
#>  [2,] 0.01354907 0.7068065 0.01354907
#>  [3,] 2.69709326 0.2818478 1.60112977
#>  [4,] 0.46913865 0.4691387 0.46913865
#>  [5,] 1.38176456 0.6086528 0.33836335
#>  [6,] 0.88325790 0.8832579 0.88325790
#>  [7,] 0.08862954 0.4412600 0.44126001
#>  [8,] 0.26242445 3.0479314 1.53610651
#>  [9,] 2.46025828 2.3552852 0.59962116
#> [10,] 0.17919648 0.1791965 0.17919648
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.9169074 0.9509832 0.47389836
#>  [2,] 0.6389964 2.0074870 1.69483357
#>  [3,] 0.4205898 0.4362379 0.55338872
#>  [4,] 1.9546905 1.2166840 0.09847026
#>  [5,] 1.0745156 0.2922691 1.31331980
#>  [6,] 0.7344728 0.7074831 0.59510825
#>  [7,] 0.7912024 4.7696579 0.85753375
#>  [8,] 0.2811077 0.1299264 0.11758480
#>  [9,] 1.7144404 2.2228947 0.32410903
#> [10,] 2.2396086 0.1305388 0.76554920
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.21317767 2.21317767 2.21317767
#>  [2,] 2.98884850 2.98884850 2.98884850
#>  [3,] 0.26010992 0.26010992 0.26010992
#>  [4,] 0.05574168 0.05574168 0.05574168
#>  [5,] 0.94836173 0.94836173 0.94836173
#>  [6,] 0.71155978 0.71155978 0.71155978
#>  [7,] 1.28588450 1.28588450 1.28588450
#>  [8,] 0.91359525 0.91359525 0.91359525
#>  [9,] 1.23674872 1.23674872 1.23674872
#> [10,] 0.43841590 0.43841590 0.43841590