Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07132710 1.42791619 0.74866700
#>  [2,] 0.09068247 0.09068247 0.09068247
#>  [3,] 1.33083966 0.91927620 0.61340054
#>  [4,] 0.51131085 0.12431699 0.51131085
#>  [5,] 0.17583481 0.29981658 0.29981658
#>  [6,] 4.04712123 3.58347303 3.58347303
#>  [7,] 0.13793830 1.67869489 0.21781165
#>  [8,] 0.71421742 0.71421742 3.81629592
#>  [9,] 1.47429704 0.43271056 1.47429704
#> [10,] 0.64430180 0.19338416 0.64430180
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]      [,2]        [,3]
#>  [1,] 0.07570043 1.6342991 0.805612736
#>  [2,] 1.23264903 1.7194599 0.244026524
#>  [3,] 0.84346889 0.7701128 0.232398948
#>  [4,] 0.35162493 0.7383153 1.310576659
#>  [5,] 0.27641427 1.3340864 1.317903911
#>  [6,] 2.03059216 2.1633001 1.281759969
#>  [7,] 0.38687026 0.2281401 0.602071033
#>  [8,] 5.77074693 1.5914870 0.002495855
#>  [9,] 2.50059605 0.2191619 0.659796226
#> [10,] 0.77944460 1.0927933 0.763737973
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 4.3431435 4.3431435 4.3431435
#>  [2,] 2.0571220 2.0571220 2.0571220
#>  [3,] 1.7498441 1.7498441 1.7498441
#>  [4,] 1.0132778 1.0132778 1.0132778
#>  [5,] 1.6354423 1.6354423 1.6354423
#>  [6,] 1.6519568 1.6519568 1.6519568
#>  [7,] 0.9319546 0.9319546 0.9319546
#>  [8,] 0.6211379 0.6211379 0.6211379
#>  [9,] 0.3090975 0.3090975 0.3090975
#> [10,] 0.8827456 0.8827456 0.8827456

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84879630 0.84153141 0.84879630
#>  [2,] 0.27876265 0.44611576 0.08251077
#>  [3,] 3.54241566 0.88432601 3.54241566
#>  [4,] 2.94035073 1.08171548 0.25037654
#>  [5,] 0.86813486 0.86813486 0.86813486
#>  [6,] 0.04471788 0.04471788 0.09945068
#>  [7,] 1.40477380 0.43507073 0.43507073
#>  [8,] 0.34681823 0.82797584 0.34681823
#>  [9,] 1.00200487 0.75124549 1.00200487
#> [10,] 0.93464678 1.20852392 1.20852392
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.9448220 0.1981827 0.10745074
#>  [2,] 2.0365767 0.1326621 0.94706174
#>  [3,] 0.9858999 0.1256956 0.20390947
#>  [4,] 1.1931798 1.9701526 0.08937071
#>  [5,] 2.5567228 0.8714429 0.15908294
#>  [6,] 1.5720231 0.4511406 0.99815287
#>  [7,] 0.2583926 0.6724293 0.23264562
#>  [8,] 1.0605146 2.2989036 0.85514916
#>  [9,] 1.2845207 1.5714210 0.06724952
#> [10,] 0.8904832 1.1961556 0.32173769
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3202702 1.3202702 1.3202702
#>  [2,] 0.0526945 0.0526945 0.0526945
#>  [3,] 0.7448976 0.7448976 0.7448976
#>  [4,] 1.2714007 1.2714007 1.2714007
#>  [5,] 0.8674635 0.8674635 0.8674635
#>  [6,] 0.6318388 0.6318388 0.6318388
#>  [7,] 0.6864750 0.6864750 0.6864750
#>  [8,] 0.6156720 0.6156720 0.6156720
#>  [9,] 0.8413464 0.8413464 0.8413464
#> [10,] 0.4140623 0.4140623 0.4140623

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.04971338 0.37470102 0.3747010
#>  [2,] 1.78510002 0.24671748 0.5425321
#>  [3,] 0.20245331 0.46473117 1.7111676
#>  [4,] 0.04845132 1.70300174 1.7030017
#>  [5,] 2.08230781 1.27848838 0.1051361
#>  [6,] 3.26157711 0.06712008 0.0279516
#>  [7,] 3.91776701 3.18281776 1.2146848
#>  [8,] 0.30660226 0.60893015 0.3066023
#>  [9,] 3.01258984 1.37628318 0.5521643
#> [10,] 1.17610325 1.57184988 1.6909368
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 2.52605468 0.138109747 0.70252368
#>  [2,] 0.05278393 0.241760174 0.71869996
#>  [3,] 1.91221645 0.195959253 0.44393248
#>  [4,] 2.75985207 0.006825189 0.61017051
#>  [5,] 0.50930307 2.045106737 0.52417366
#>  [6,] 0.18606819 2.279078494 0.90533003
#>  [7,] 0.69168179 0.571645453 0.03363697
#>  [8,] 0.04864679 1.142501364 1.21053604
#>  [9,] 1.54910449 0.611222841 0.08650359
#> [10,] 0.68519958 0.181912745 2.50098074
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 3.5591888 3.5591888 3.5591888
#>  [2,] 0.3765026 0.3765026 0.3765026
#>  [3,] 3.0801485 3.0801485 3.0801485
#>  [4,] 2.4214955 2.4214955 2.4214955
#>  [5,] 1.7766114 1.7766114 1.7766114
#>  [6,] 2.4668225 2.4668225 2.4668225
#>  [7,] 0.9333240 0.9333240 0.9333240
#>  [8,] 0.5229361 0.5229361 0.5229361
#>  [9,] 0.8108746 0.8108746 0.8108746
#> [10,] 1.0063746 1.0063746 1.0063746