Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.8006275 2.02603625 0.2330134
#>  [2,] 0.4476923 0.48017283 0.4801728
#>  [3,] 1.0033278 0.49434739 0.1332174
#>  [4,] 0.5428004 0.38098816 0.5428004
#>  [5,] 0.1878339 0.56442015 0.6510687
#>  [6,] 0.2946568 0.69170012 0.2946568
#>  [7,] 0.9015612 1.53043042 1.5304304
#>  [8,] 2.7577939 2.46371443 1.0359932
#>  [9,] 0.8768334 0.08302871 1.3132333
#> [10,] 0.2094394 1.65121566 1.6512157
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.11487789 0.20008763 1.24663129
#>  [2,] 1.91918689 2.81746934 1.70586363
#>  [3,] 0.01779804 5.58021262 0.28793068
#>  [4,] 1.48894795 1.51043345 0.54314661
#>  [5,] 0.16239161 1.08956049 2.06899246
#>  [6,] 0.77349868 3.05814691 0.21801311
#>  [7,] 2.12716572 0.01770344 0.06386404
#>  [8,] 0.71479829 1.01787587 0.07835613
#>  [9,] 0.92811108 0.50273675 0.28034397
#> [10,] 0.93229424 1.13973642 0.67075059
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.09696289 1.09696289 1.09696289
#>  [2,] 0.01080994 0.01080994 0.01080994
#>  [3,] 1.68793060 1.68793060 1.68793060
#>  [4,] 4.06749891 4.06749891 4.06749891
#>  [5,] 0.29008221 0.29008221 0.29008221
#>  [6,] 0.52000387 0.52000387 0.52000387
#>  [7,] 0.67083965 0.67083965 0.67083965
#>  [8,] 0.08314637 0.08314637 0.08314637
#>  [9,] 1.08778338 1.08778338 1.08778338
#> [10,] 1.73429384 1.73429384 1.73429384

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.95283481 0.4711868 3.1585004
#>  [2,] 0.01422628 0.3767208 0.2912219
#>  [3,] 0.36892061 0.3478707 0.1835459
#>  [4,] 0.92537196 3.8170987 2.1737336
#>  [5,] 0.80750952 0.8075095 1.2327438
#>  [6,] 0.79396558 0.7939656 0.7939656
#>  [7,] 0.14569955 0.1456996 0.1456996
#>  [8,] 1.46146404 0.4034090 1.4614640
#>  [9,] 0.03904438 1.5204038 0.9226741
#> [10,] 0.83613509 0.8361351 0.8361351
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.57876495 0.93069326 0.41729411
#>  [2,] 0.07070152 1.22345129 0.02718375
#>  [3,] 0.70221536 0.55371926 0.55264817
#>  [4,] 1.48006493 2.36709523 0.40472785
#>  [5,] 1.89418239 1.41118738 0.51646870
#>  [6,] 2.42853155 1.11797853 0.16658437
#>  [7,] 0.67688009 0.08662434 3.88453696
#>  [8,] 4.20418899 1.31105787 0.07123370
#>  [9,] 1.01740087 1.23775494 0.28445928
#> [10,] 0.41328271 0.46287602 0.43008569
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09020151 0.09020151 0.09020151
#>  [2,] 0.84545545 0.84545545 0.84545545
#>  [3,] 0.33488334 0.33488334 0.33488334
#>  [4,] 0.06860975 0.06860975 0.06860975
#>  [5,] 0.07838095 0.07838095 0.07838095
#>  [6,] 0.03421787 0.03421787 0.03421787
#>  [7,] 0.92499732 0.92499732 0.92499732
#>  [8,] 1.19169290 1.19169290 1.19169290
#>  [9,] 0.95912681 0.95912681 0.95912681
#> [10,] 0.40158515 0.40158515 0.40158515

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.22718167 0.44525181 0.44525181
#>  [2,] 1.68073794 0.71594030 0.21582117
#>  [3,] 0.93165424 0.93165424 0.93165424
#>  [4,] 1.15604325 1.15604325 1.15604325
#>  [5,] 0.29729630 0.29729630 0.29729630
#>  [6,] 1.52810367 1.52810367 1.52810367
#>  [7,] 0.56150215 1.00665514 1.87260082
#>  [8,] 0.02751885 0.02751885 0.02751885
#>  [9,] 1.61750357 0.69973013 0.68825673
#> [10,] 0.78993252 0.78993252 0.78993252
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 2.6886118 0.0425595 3.19246913
#>  [2,] 0.2468871 1.4386837 0.54191522
#>  [3,] 0.6848429 0.3063247 0.85573032
#>  [4,] 1.4328375 2.9919598 0.97799285
#>  [5,] 0.6648623 1.3046987 0.04418445
#>  [6,] 0.2016367 2.3158024 0.03275788
#>  [7,] 2.4495765 0.5989868 4.68002503
#>  [8,] 0.3219650 0.0609751 0.14963699
#>  [9,] 0.2110932 1.6121359 0.22315132
#> [10,] 1.0665893 0.4414501 0.41484299
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.005334848 0.005334848 0.005334848
#>  [2,] 2.131928133 2.131928133 2.131928133
#>  [3,] 0.693259042 0.693259042 0.693259042
#>  [4,] 0.736132775 0.736132775 0.736132775
#>  [5,] 1.262141852 1.262141852 1.262141852
#>  [6,] 1.327315440 1.327315440 1.327315440
#>  [7,] 0.063889788 0.063889788 0.063889788
#>  [8,] 0.168690098 0.168690098 0.168690098
#>  [9,] 0.520691216 0.520691216 0.520691216
#> [10,] 0.082897230 0.082897230 0.082897230