Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>              [,1]       [,2]       [,3]
#>  [1,] 0.743557268 1.21465922 0.09518423
#>  [2,] 0.596715929 0.59671593 0.32401021
#>  [3,] 5.265141646 0.02300012 0.26042776
#>  [4,] 0.753946009 0.25311262 0.82962228
#>  [5,] 0.314950353 0.31495035 0.31495035
#>  [6,] 1.239595839 0.95321514 0.33333559
#>  [7,] 2.829315827 1.82886435 0.01518806
#>  [8,] 0.104302513 1.00796304 1.75220370
#>  [9,] 2.515386863 1.64058371 1.70029025
#> [10,] 0.002606246 1.27940670 1.35096917
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>            [,1]      [,2]       [,3]
#>  [1,] 1.4268968 0.1190370 2.63465930
#>  [2,] 0.8745607 3.7411576 1.26712548
#>  [3,] 0.2514691 0.6132234 1.16584613
#>  [4,] 5.9371980 1.9655323 3.09191106
#>  [5,] 1.2310482 0.6304577 0.32103804
#>  [6,] 0.8785980 0.3133107 2.27684105
#>  [7,] 0.7714461 2.7067545 0.03260316
#>  [8,] 0.6416997 0.7150591 0.84499230
#>  [9,] 0.7497388 1.6316350 0.20710108
#> [10,] 0.3796644 0.2730875 0.73971773
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21121759 0.21121759 0.21121759
#>  [2,] 0.03387643 0.03387643 0.03387643
#>  [3,] 0.42716733 0.42716733 0.42716733
#>  [4,] 2.06393080 2.06393080 2.06393080
#>  [5,] 1.50678599 1.50678599 1.50678599
#>  [6,] 0.65434526 0.65434526 0.65434526
#>  [7,] 1.77007773 1.77007773 1.77007773
#>  [8,] 0.19639190 0.19639190 0.19639190
#>  [9,] 0.63624085 0.63624085 0.63624085
#> [10,] 0.44660102 0.44660102 0.44660102

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.0130723 0.01307230 0.01307230
#>  [2,] 0.4692227 0.46922267 1.20747627
#>  [3,] 0.4386094 0.43860943 0.43860943
#>  [4,] 0.2863835 0.41543498 0.28638354
#>  [5,] 3.9277737 3.13406466 0.03620016
#>  [6,] 1.7519034 1.75190342 1.75190342
#>  [7,] 0.4567640 0.09854018 0.45676399
#>  [8,] 0.1212160 0.12121596 0.12121596
#>  [9,] 1.2062101 0.24229276 0.24229276
#> [10,] 6.6508670 4.34112486 3.84372967
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>            [,1]      [,2]       [,3]
#>  [1,] 3.7194046 0.6863766 1.77602020
#>  [2,] 0.7078609 0.7732670 0.51506537
#>  [3,] 1.0015626 0.4416403 1.44271224
#>  [4,] 0.6209279 0.3772850 0.44522070
#>  [5,] 0.4806551 0.3404420 0.54232673
#>  [6,] 0.3795646 2.6962903 0.24399888
#>  [7,] 0.5321688 0.9715679 1.84737374
#>  [8,] 0.1759405 0.3008692 1.38762114
#>  [9,] 0.4655567 1.2071981 0.07905209
#> [10,] 0.4170242 0.2373415 0.05251248
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5488749 1.5488749 1.5488749
#>  [2,] 0.1564271 0.1564271 0.1564271
#>  [3,] 0.7665288 0.7665288 0.7665288
#>  [4,] 3.2092532 3.2092532 3.2092532
#>  [5,] 1.2077471 1.2077471 1.2077471
#>  [6,] 0.1029387 0.1029387 0.1029387
#>  [7,] 3.2374715 3.2374715 3.2374715
#>  [8,] 1.4480081 1.4480081 1.4480081
#>  [9,] 1.8542464 1.8542464 1.8542464
#> [10,] 0.9702451 0.9702451 0.9702451

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3107373 0.31073734 0.31073734
#>  [2,] 0.1221733 0.04927565 0.12217329
#>  [3,] 0.4945307 1.04126550 1.04126550
#>  [4,] 2.1357714 0.11899336 1.27216470
#>  [5,] 2.0112538 0.66143382 0.28034667
#>  [6,] 1.8113010 1.47490838 1.44005341
#>  [7,] 0.5074258 0.50742582 1.85765430
#>  [8,] 0.4619162 0.01567832 0.46191617
#>  [9,] 5.5302187 5.53021872 2.42813608
#> [10,] 0.0771986 0.07719860 0.01148501
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>              [,1]       [,2]      [,3]
#>  [1,] 0.417002154 0.03604627 0.4596523
#>  [2,] 0.526978934 0.73041023 1.3964598
#>  [3,] 1.225644596 0.73754549 0.8575663
#>  [4,] 1.347130044 0.68794666 0.1382868
#>  [5,] 0.009647696 0.41953782 1.4154975
#>  [6,] 1.247943329 1.13511409 1.0860708
#>  [7,] 0.148708461 0.64221146 2.5822547
#>  [8,] 2.057843922 0.79282352 0.5298258
#>  [9,] 0.200835653 1.40006813 2.0824356
#> [10,] 0.562180240 0.70952150 0.9309263
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06754695 0.06754695 0.06754695
#>  [2,] 0.76168856 0.76168856 0.76168856
#>  [3,] 0.94647695 0.94647695 0.94647695
#>  [4,] 0.75229350 0.75229350 0.75229350
#>  [5,] 3.41360018 3.41360018 3.41360018
#>  [6,] 0.86357831 0.86357831 0.86357831
#>  [7,] 1.37650925 1.37650925 1.37650925
#>  [8,] 0.50577880 0.50577880 0.50577880
#>  [9,] 0.21720710 0.21720710 0.21720710
#> [10,] 2.45633058 2.45633058 2.45633058