Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]      [,3]
#>  [1,] 1.91574568 0.63147915 0.8402114
#>  [2,] 2.97843576 0.34989521 1.2181080
#>  [3,] 0.06321985 0.08331033 4.4540728
#>  [4,] 1.62882490 1.51710459 1.5171046
#>  [5,] 0.81053269 0.81053269 0.8105327
#>  [6,] 0.27814151 0.27814151 0.2781415
#>  [7,] 1.07393447 1.07393447 0.1711430
#>  [8,] 1.82264456 0.67466117 1.8226446
#>  [9,] 0.57771382 0.57771382 1.1920909
#> [10,] 0.25129515 0.38029301 0.2512951
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 2.19372853 0.81820244 2.91573239
#>  [2,] 3.72045857 0.81323187 0.92892877
#>  [3,] 0.22479202 0.43350328 0.04797820
#>  [4,] 0.04958342 0.35272593 2.11865253
#>  [5,] 0.33161603 0.78016925 1.24386153
#>  [6,] 1.34622656 3.67325016 0.85957761
#>  [7,] 1.29810297 1.87992552 2.26542030
#>  [8,] 0.51504467 0.23669966 0.81940658
#>  [9,] 0.61946821 0.07969184 0.09538103
#> [10,] 0.24801913 0.23658298 0.05308767
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.17042698 1.17042698 1.17042698
#>  [2,] 1.88079083 1.88079083 1.88079083
#>  [3,] 0.40087048 0.40087048 0.40087048
#>  [4,] 1.21523061 1.21523061 1.21523061
#>  [5,] 0.04267716 0.04267716 0.04267716
#>  [6,] 0.01795197 0.01795197 0.01795197
#>  [7,] 0.80548060 0.80548060 0.80548060
#>  [8,] 1.31625056 1.31625056 1.31625056
#>  [9,] 0.14766399 0.14766399 0.14766399
#> [10,] 2.00637036 2.00637036 2.00637036

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5675064 0.5675064 0.2613325
#>  [2,] 2.3628038 1.8721056 3.8324526
#>  [3,] 0.3105511 0.3105511 0.3105511
#>  [4,] 0.4816685 0.8626386 0.2026987
#>  [5,] 0.3451092 0.3451092 0.3451092
#>  [6,] 1.4321962 0.6605746 1.0028064
#>  [7,] 1.6445587 1.1268061 1.1268061
#>  [8,] 0.9871734 0.1860931 0.9871734
#>  [9,] 0.2848356 0.2848356 0.2229842
#> [10,] 1.0700808 1.0700808 1.1058964
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>            [,1]       [,2]        [,3]
#>  [1,] 1.1417019 0.60192714 0.007078525
#>  [2,] 0.5060432 2.35811639 0.490891478
#>  [3,] 1.0978726 0.08256637 0.259483562
#>  [4,] 1.2249615 0.60894024 0.519289597
#>  [5,] 0.8219217 1.72665830 5.277008343
#>  [6,] 0.6277124 0.08074320 0.371533989
#>  [7,] 0.3931046 1.75195089 1.016929414
#>  [8,] 0.2852436 2.13926623 2.621621626
#>  [9,] 1.2267496 0.46211865 0.188345064
#> [10,] 1.6255587 1.08432211 1.462587545
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2152534 0.2152534 0.2152534
#>  [2,] 1.0640906 1.0640906 1.0640906
#>  [3,] 0.7847122 0.7847122 0.7847122
#>  [4,] 5.2474691 5.2474691 5.2474691
#>  [5,] 2.9263869 2.9263869 2.9263869
#>  [6,] 0.5645700 0.5645700 0.5645700
#>  [7,] 0.8935550 0.8935550 0.8935550
#>  [8,] 0.2720553 0.2720553 0.2720553
#>  [9,] 0.3348159 0.3348159 0.3348159
#> [10,] 0.9724681 0.9724681 0.9724681

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1949570 0.1023448 0.2261147
#>  [2,] 0.2231226 0.3298126 0.3776847
#>  [3,] 1.2349476 0.3297421 1.7098538
#>  [4,] 0.2339178 0.2339178 0.2339178
#>  [5,] 1.0717857 0.9858491 0.3584959
#>  [6,] 0.3424026 1.6809396 1.6809396
#>  [7,] 0.2689254 0.2689254 0.2689254
#>  [8,] 1.5131186 4.1125191 0.4693709
#>  [9,] 0.4953869 0.4953869 0.4953869
#> [10,] 0.1873238 0.1873238 0.1873238
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.60781126 2.01756063 2.9519287
#>  [2,] 0.78307051 0.51841418 0.1598147
#>  [3,] 5.11900717 0.21054404 0.7864646
#>  [4,] 0.01811362 0.06253545 1.3660061
#>  [5,] 0.51517199 1.06761830 0.8666666
#>  [6,] 1.00818243 0.06801344 0.0426063
#>  [7,] 1.72897950 0.24725696 0.3649072
#>  [8,] 0.60049782 1.08609215 5.0030355
#>  [9,] 0.12903208 2.66692999 0.3215650
#> [10,] 3.08924434 0.80516042 0.3116466
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.87556133 0.87556133 0.87556133
#>  [2,] 0.01598692 0.01598692 0.01598692
#>  [3,] 0.47958957 0.47958957 0.47958957
#>  [4,] 0.91302761 0.91302761 0.91302761
#>  [5,] 0.21946247 0.21946247 0.21946247
#>  [6,] 1.17432284 1.17432284 1.17432284
#>  [7,] 0.45513203 0.45513203 0.45513203
#>  [8,] 3.13648280 3.13648280 3.13648280
#>  [9,] 0.34518232 0.34518232 0.34518232
#> [10,] 2.30583836 2.30583836 2.30583836