Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.60529085 3.07770981 1.02875180
#>  [2,] 1.78546148 3.02408068 1.78546148
#>  [3,] 0.64154880 0.64154880 0.64154880
#>  [4,] 1.34414147 4.04443957 1.63203219
#>  [5,] 1.60301155 2.32483794 2.94782721
#>  [6,] 2.23525453 1.74254093 1.20178435
#>  [7,] 0.36277570 0.07139836 0.07139836
#>  [8,] 0.08725499 1.64191935 1.64191935
#>  [9,] 0.30115233 1.88922429 2.60155269
#> [10,] 0.10324488 0.10324488 0.10324488
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1))
#>            [,1]      [,2]         [,3]
#>  [1,] 0.7048605 2.5973511 1.686175e+00
#>  [2,] 0.5998165 0.8561126 1.271186e+00
#>  [3,] 1.7841613 0.4724216 5.090837e-01
#>  [4,] 1.1761410 0.9265426 1.533521e+00
#>  [5,] 0.4967938 5.0414097 1.857342e+00
#>  [6,] 1.3681333 0.4092560 1.884725e+00
#>  [7,] 0.3516439 0.9273949 3.003500e-01
#>  [8,] 0.4857643 1.6038542 3.591742e-01
#>  [9,] 3.1304954 1.9614541 7.193567e-05
#> [10,] 0.2655322 0.2030091 2.762151e+00
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.03154979 1.03154979 1.03154979
#>  [2,] 0.24010448 0.24010448 0.24010448
#>  [3,] 2.21467028 2.21467028 2.21467028
#>  [4,] 0.29575483 0.29575483 0.29575483
#>  [5,] 5.14732866 5.14732866 5.14732866
#>  [6,] 1.30446664 1.30446664 1.30446664
#>  [7,] 1.32733444 1.32733444 1.32733444
#>  [8,] 0.29771318 0.29771318 0.29771318
#>  [9,] 0.09491203 0.09491203 0.09491203
#> [10,] 0.17882509 0.17882509 0.17882509

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.33677638 0.3367764 0.3367764
#>  [2,] 0.67121521 0.6712152 0.6712152
#>  [3,] 0.85948193 0.8594819 2.5296376
#>  [4,] 0.85145441 3.2834063 0.2222569
#>  [5,] 1.88846527 2.2965396 0.7417268
#>  [6,] 0.20206660 1.1593857 0.5299341
#>  [7,] 2.19881597 1.5459265 2.1988160
#>  [8,] 0.09463519 0.9483721 0.1087106
#>  [9,] 1.10745058 0.8508959 1.1074506
#> [10,] 0.87162376 0.8716238 0.4146740
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM")
#>              [,1]      [,2]       [,3]
#>  [1,] 1.171368686 0.5569533 0.05278705
#>  [2,] 0.003161547 0.4303067 0.11099502
#>  [3,] 0.341519775 1.4376279 0.66110563
#>  [4,] 0.553923332 0.2219836 0.60492007
#>  [5,] 1.635457389 0.8981412 0.99814993
#>  [6,] 0.687745846 0.1199130 1.61732268
#>  [7,] 0.177721891 0.5081820 0.05091197
#>  [8,] 0.050980424 0.5955293 0.28251212
#>  [9,] 1.352231093 0.9131305 0.57736997
#> [10,] 1.104992767 0.2170265 0.14317609
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.50192038 1.50192038 1.50192038
#>  [2,] 1.88910073 1.88910073 1.88910073
#>  [3,] 1.90309421 1.90309421 1.90309421
#>  [4,] 0.32633489 0.32633489 0.32633489
#>  [5,] 0.07101626 0.07101626 0.07101626
#>  [6,] 0.24139950 0.24139950 0.24139950
#>  [7,] 0.96453409 0.96453409 0.96453409
#>  [8,] 0.35115557 0.35115557 0.35115557
#>  [9,] 1.05615883 1.05615883 1.05615883
#> [10,] 3.11955955 3.11955955 3.11955955

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4435824 0.4435824 0.44358245
#>  [2,] 2.0070470 0.7827388 0.23132896
#>  [3,] 0.5944471 0.5944471 0.01383497
#>  [4,] 2.2500752 4.9692837 0.62711764
#>  [5,] 0.7538751 1.2208600 2.73341236
#>  [6,] 0.3165111 2.3512622 4.07697716
#>  [7,] 1.3939420 0.2246526 1.20683590
#>  [8,] 0.5882063 2.7532439 0.58638467
#>  [9,] 0.4311112 0.4311112 0.43111122
#> [10,] 2.4602307 0.5034221 2.46023066
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9470584 0.5489734 0.36609401
#>  [2,] 1.6920380 0.2152973 0.75120039
#>  [3,] 1.8241327 0.1930355 1.93183726
#>  [4,] 0.2054963 0.9191453 1.43042216
#>  [5,] 0.4682803 1.9806972 0.94300117
#>  [6,] 0.3200544 0.6594477 1.03026564
#>  [7,] 1.1956992 0.4871992 0.98719286
#>  [8,] 0.8762022 1.0355150 0.03233961
#>  [9,] 0.7557547 1.2547101 1.56134100
#> [10,] 0.2694071 1.2053248 3.30171044
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.266851880 0.266851880 0.266851880
#>  [2,] 0.139322366 0.139322366 0.139322366
#>  [3,] 0.131623053 0.131623053 0.131623053
#>  [4,] 0.858561370 0.858561370 0.858561370
#>  [5,] 0.007212656 0.007212656 0.007212656
#>  [6,] 1.834185541 1.834185541 1.834185541
#>  [7,] 0.527198181 0.527198181 0.527198181
#>  [8,] 0.177736669 0.177736669 0.177736669
#>  [9,] 0.375888867 0.375888867 0.375888867
#> [10,] 1.630675297 1.630675297 1.630675297