Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05807002 0.05807002 0.05807002
#>  [2,] 0.13507020 0.13507020 0.13507020
#>  [3,] 0.22662991 0.86199428 0.86199428
#>  [4,] 0.14912552 0.14912552 0.14912552
#>  [5,] 0.66043707 0.67552088 0.66043707
#>  [6,] 0.24628748 0.50667041 0.53751812
#>  [7,] 0.06200684 0.06200684 0.04393674
#>  [8,] 0.91369220 0.91369220 0.38749454
#>  [9,] 0.78924219 0.78924219 0.78924219
#> [10,] 0.33482931 0.33482931 1.71296046
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2789890 0.8218194 1.43680460
#>  [2,] 0.2091470 1.4132513 1.15645763
#>  [3,] 4.4787404 0.2342501 0.20351741
#>  [4,] 3.2532646 3.0267099 0.86978949
#>  [5,] 0.6371994 0.2842573 0.02671364
#>  [6,] 0.9173684 0.9320403 1.22341277
#>  [7,] 0.7835603 1.4242171 1.56899966
#>  [8,] 1.0450095 0.5604690 0.85135152
#>  [9,] 3.6174406 0.7891777 1.83620873
#> [10,] 1.0869715 0.2545839 0.83801449
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15230312 0.15230312 0.15230312
#>  [2,] 2.00631492 2.00631492 2.00631492
#>  [3,] 1.12834999 1.12834999 1.12834999
#>  [4,] 0.76094727 0.76094727 0.76094727
#>  [5,] 3.51902833 3.51902833 3.51902833
#>  [6,] 0.56343120 0.56343120 0.56343120
#>  [7,] 1.01967302 1.01967302 1.01967302
#>  [8,] 0.44202634 0.44202634 0.44202634
#>  [9,] 0.05050234 0.05050234 0.05050234
#> [10,] 1.31575274 1.31575274 1.31575274

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5281126 1.48119174 0.33613914
#>  [2,] 0.2093256 0.46788768 0.20932555
#>  [3,] 3.7668231 0.22505217 3.76682307
#>  [4,] 0.1507409 0.15074094 0.15074094
#>  [5,] 0.2758862 1.76437653 0.09452116
#>  [6,] 0.1463998 0.03109344 0.05871276
#>  [7,] 1.7494044 0.08508589 0.72917461
#>  [8,] 0.1739391 0.17393910 0.17393910
#>  [9,] 0.4471017 0.73618895 0.85762687
#> [10,] 1.0191695 0.36724937 0.16382490
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>              [,1]       [,2]       [,3]
#>  [1,] 1.214424536 0.06719777 0.40236478
#>  [2,] 0.286823173 3.99690589 0.34452176
#>  [3,] 0.001238007 2.59314185 0.05148465
#>  [4,] 1.434186591 0.37295139 0.37449700
#>  [5,] 0.331261494 0.89555245 0.62798416
#>  [6,] 0.174268467 1.68475235 2.73148790
#>  [7,] 0.475310484 0.16610513 0.52444641
#>  [8,] 2.992570558 0.91928964 0.08508282
#>  [9,] 0.929041611 0.88243588 0.20063790
#> [10,] 0.148944195 3.01813556 0.04066414
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.68352300 3.68352300 3.68352300
#>  [2,] 0.87891849 0.87891849 0.87891849
#>  [3,] 0.80589366 0.80589366 0.80589366
#>  [4,] 1.79618835 1.79618835 1.79618835
#>  [5,] 0.07665642 0.07665642 0.07665642
#>  [6,] 0.80263655 0.80263655 0.80263655
#>  [7,] 0.94357766 0.94357766 0.94357766
#>  [8,] 0.09956454 0.09956454 0.09956454
#>  [9,] 3.65212632 3.65212632 3.65212632
#> [10,] 0.35467884 0.35467884 0.35467884

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8749616 3.0286287 1.10108900
#>  [2,] 1.5718398 0.3015148 0.77263144
#>  [3,] 0.8416060 0.6350483 0.63504828
#>  [4,] 1.4799784 3.1131400 0.44265351
#>  [5,] 1.3958601 2.8696681 3.11054770
#>  [6,] 0.2857787 0.3580929 0.09869569
#>  [7,] 4.9166944 0.6345417 0.63454168
#>  [8,] 1.6140087 1.0241176 0.25220376
#>  [9,] 2.6074555 0.4207639 0.42076392
#> [10,] 0.6139366 0.6139366 0.25425595
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04271528 0.09743509 0.12876383
#>  [2,] 0.71938199 2.13562547 1.54346069
#>  [3,] 0.77657084 1.14399967 0.54352759
#>  [4,] 1.07547466 0.65214172 3.32886751
#>  [5,] 0.66725446 0.36809920 1.28368168
#>  [6,] 0.41126931 0.14905156 0.65121210
#>  [7,] 6.27648185 1.47788206 0.02202653
#>  [8,] 0.44056132 0.05073758 0.30169405
#>  [9,] 0.19321929 1.32901866 0.50143619
#> [10,] 0.88227215 0.03582561 1.77599146
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.2737554 2.2737554 2.2737554
#>  [2,] 0.0185998 0.0185998 0.0185998
#>  [3,] 0.9026229 0.9026229 0.9026229
#>  [4,] 3.9318448 3.9318448 3.9318448
#>  [5,] 1.9657515 1.9657515 1.9657515
#>  [6,] 0.1102076 0.1102076 0.1102076
#>  [7,] 3.3871400 3.3871400 3.3871400
#>  [8,] 0.1048714 0.1048714 0.1048714
#>  [9,] 0.3481090 0.3481090 0.3481090
#> [10,] 1.0597164 1.0597164 1.0597164