Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

Calculation of the shock-size arrival intensities

The exchangeable shock-size arrival intensities of a constant Bernstein function with triplet \((a, 0, 0)\) are $$ \eta_{i} = \binom{d}{i} a 1_{\{ i = d \}} , \quad i \in {\{ 1 , \ldots , d \}} . $$

The exchangeable shock-size arrival intensities of a linear Bernstein function with triplet \((0, b, 0)\) are $$ \eta_{i} = \binom{d}{i} b 1_{\{ i = 1 \}} , \quad i \in {\{ 1 , \ldots , d \}} . $$

The exchangeable shock-size arrival intensities of a Bernstein function with Lévy triple \((0, 0, \nu)\) are $$ \eta_{i} = \int_{0}^{\infty} \binom{d}{i} {(1 - e^{-u})}^{i} e^{- u {(d-i)}} \nu{(du)} , \quad i \in {\{ 1, \ldots, d \}} $$

The exchangeable shock-size arrival intensities of a complete Bernstein function with Stieltjes triple \((0, 0, \sigma)\) are $$ \eta_{i} = \int_{0}^{\infty} \binom{d}{i} u B{(1 + i, d - i + u)} \sigma{(du)} , \quad i \in {\{ 1, \ldots, d \}} $$

The exchangeable shock-size arrival intensities of convex combinations of Bernstein functions are the corresponding convex combinations of their individual exchangeable shock-size arrival intensities with the same coefficients.

See (Sloot 2022) for the background, and detailed notes on numerical approximations of these integrals.

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Sloot H (2022). “Implementing Markovian models for extendible Marshall–Olkin distributions.” Dependence Modeling, 10(1), 308--343. doi:10.1515/demo-2022-0151 , https://doi.org/doi:10.1515/demo-2022-0151.

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2803955 0.2803955 0.2803955
#>  [2,] 1.6225232 0.5497167 1.6225232
#>  [3,] 0.3398122 1.8139854 1.8139854
#>  [4,] 0.2780088 0.2780088 0.2780088
#>  [5,] 0.1123318 0.5376350 0.5842482
#>  [6,] 0.7282013 2.4388738 0.7379427
#>  [7,] 3.9002625 1.8898198 1.8898198
#>  [8,] 4.7017037 2.4008320 0.1392069
#>  [9,] 0.2115317 2.3774982 0.4278534
#> [10,] 0.5061890 0.5061890 0.5061890
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.76988107 0.90297283 1.75222793
#>  [2,] 1.18719605 0.05613957 0.04143088
#>  [3,] 2.23495045 0.80584824 0.89927103
#>  [4,] 0.32798448 1.25129797 0.19821565
#>  [5,] 0.15860899 0.54712411 0.38081558
#>  [6,] 1.07169279 0.02945748 1.53146735
#>  [7,] 0.22093089 0.06170399 0.43208323
#>  [8,] 0.65919015 0.04955331 1.72310566
#>  [9,] 0.35885331 0.73695324 0.44483540
#> [10,] 0.08931908 0.28831518 1.25441962
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.8938342 2.8938342 2.8938342
#>  [2,] 1.6456940 1.6456940 1.6456940
#>  [3,] 0.3331875 0.3331875 0.3331875
#>  [4,] 1.1828322 1.1828322 1.1828322
#>  [5,] 0.5237272 0.5237272 0.5237272
#>  [6,] 1.7504132 1.7504132 1.7504132
#>  [7,] 0.9190773 0.9190773 0.9190773
#>  [8,] 0.3711171 0.3711171 0.3711171
#>  [9,] 0.4172682 0.4172682 0.4172682
#> [10,] 2.0289819 2.0289819 2.0289819

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.03471777 0.86174095 0.86174095
#>  [2,] 0.07268309 0.07268309 0.39990319
#>  [3,] 0.31024519 0.31024519 0.31024519
#>  [4,] 1.05636383 0.06757005 0.06757005
#>  [5,] 0.16227534 0.24964548 0.24964548
#>  [6,] 1.99802860 0.25136324 0.25136324
#>  [7,] 0.38687923 0.38687923 0.38687923
#>  [8,] 3.98266057 2.06529773 0.65989997
#>  [9,] 1.91480505 2.12774465 1.30620079
#> [10,] 0.12159583 0.63963588 2.91366652
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 3.5061041 0.64762009 0.37399297
#>  [2,] 0.4292600 0.41143352 0.12162760
#>  [3,] 1.1404568 0.21273955 0.09323977
#>  [4,] 1.7017278 0.64203649 1.42370456
#>  [5,] 2.1742562 1.28670651 1.74625406
#>  [6,] 0.1525817 1.13065009 0.95303139
#>  [7,] 0.1587143 0.08472629 0.12599636
#>  [8,] 1.6159992 0.07525890 1.90471265
#>  [9,] 0.3733149 0.92345342 0.93325952
#> [10,] 3.4930417 1.01347601 2.00853188
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6240285 0.6240285 0.6240285
#>  [2,] 0.3864394 0.3864394 0.3864394
#>  [3,] 3.3549797 3.3549797 3.3549797
#>  [4,] 1.7281929 1.7281929 1.7281929
#>  [5,] 3.5406422 3.5406422 3.5406422
#>  [6,] 2.7970685 2.7970685 2.7970685
#>  [7,] 1.0758359 1.0758359 1.0758359
#>  [8,] 0.9012577 0.9012577 0.9012577
#>  [9,] 0.7286548 0.7286548 0.7286548
#> [10,] 1.0252282 1.0252282 1.0252282

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.50699700 1.2907323 1.88721911
#>  [2,] 0.08679736 0.3749210 2.67502019
#>  [3,] 0.43835296 0.9723530 0.97235300
#>  [4,] 0.13847086 0.2582989 0.16211764
#>  [5,] 1.55675035 2.5780045 1.55675035
#>  [6,] 1.84926888 0.5571417 1.84926888
#>  [7,] 0.44061421 0.4406142 0.01936539
#>  [8,] 0.53682783 0.1625573 0.42462409
#>  [9,] 1.43068463 3.1334280 1.43068463
#> [10,] 1.75566956 0.6346167 0.63461670
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2159359 0.25510351 0.6117571
#>  [2,] 0.5384507 0.01617066 0.3358434
#>  [3,] 2.1687818 0.48153244 2.0783484
#>  [4,] 0.1698161 0.57529354 2.8301581
#>  [5,] 1.8407087 1.07318544 0.3977137
#>  [6,] 1.9462738 1.28824593 0.3900389
#>  [7,] 2.2890779 0.33153577 2.2133841
#>  [8,] 1.9797259 0.24243429 1.2507619
#>  [9,] 1.2451751 2.74131959 0.2015741
#> [10,] 0.1588986 0.26062562 2.1581168
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1444033 0.1444033 0.1444033
#>  [2,] 0.5319950 0.5319950 0.5319950
#>  [3,] 1.0829017 1.0829017 1.0829017
#>  [4,] 0.8465788 0.8465788 0.8465788
#>  [5,] 1.2360087 1.2360087 1.2360087
#>  [6,] 1.3186094 1.3186094 1.3186094
#>  [7,] 0.1079567 0.1079567 0.1079567
#>  [8,] 1.8837483 1.8837483 1.8837483
#>  [9,] 0.6249595 0.6249595 0.6249595
#> [10,] 3.7551706 3.7551706 3.7551706