Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.18470496 0.8259091 0.82590905
#>  [2,] 1.11431346 0.9694259 2.23022804
#>  [3,] 0.28408819 0.9153908 0.04588380
#>  [4,] 2.82823204 1.9048839 1.90488389
#>  [5,] 0.28113691 2.0693934 0.26695178
#>  [6,] 3.08209669 0.3952805 1.48485715
#>  [7,] 0.42853217 1.2303428 1.23034277
#>  [8,] 1.87096394 0.9107382 1.87096394
#>  [9,] 0.13036182 0.6669819 1.17940472
#> [10,] 0.09267647 1.6224080 0.09267647
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.44258628 0.09017764 1.2280056
#>  [2,] 0.01180354 0.50498311 0.4237390
#>  [3,] 1.24800106 0.31792840 1.0148471
#>  [4,] 3.41057593 1.44499877 1.2545123
#>  [5,] 0.13924829 1.31900751 0.3822840
#>  [6,] 0.17938884 0.50749398 1.9819168
#>  [7,] 1.31291901 0.08393576 0.3415356
#>  [8,] 4.09783922 0.09322118 0.1742064
#>  [9,] 0.20291500 0.57011776 0.4248982
#> [10,] 0.93319008 0.33921121 0.5776025
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55005045 0.55005045 0.55005045
#>  [2,] 0.01987800 0.01987800 0.01987800
#>  [3,] 2.05481126 2.05481126 2.05481126
#>  [4,] 0.06812646 0.06812646 0.06812646
#>  [5,] 0.02470566 0.02470566 0.02470566
#>  [6,] 2.24520558 2.24520558 2.24520558
#>  [7,] 0.15923101 0.15923101 0.15923101
#>  [8,] 1.02757122 1.02757122 1.02757122
#>  [9,] 0.16788103 0.16788103 0.16788103
#> [10,] 0.30467424 0.30467424 0.30467424

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 2.56417384 2.564173839 2.56417384
#>  [2,] 0.19651772 0.003144742 0.19651772
#>  [3,] 0.87771297 0.125512951 0.03694926
#>  [4,] 0.07554558 1.836109283 0.72604614
#>  [5,] 1.30421289 1.304212891 0.50214533
#>  [6,] 0.13376765 0.160015744 0.13376765
#>  [7,] 2.22472280 1.827386077 2.79336699
#>  [8,] 1.27942943 0.165679354 0.16567935
#>  [9,] 0.20895336 0.018984949 0.09964746
#> [10,] 0.05895449 0.174074977 1.92353770
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08430583 3.59256228 1.33622096
#>  [2,] 0.22525025 0.15574763 0.77048251
#>  [3,] 0.55915787 1.47534868 1.36340608
#>  [4,] 0.34048381 0.66272098 0.17679197
#>  [5,] 0.67617937 0.25289318 0.65371813
#>  [6,] 3.02811194 0.25920261 0.02349767
#>  [7,] 0.62365464 0.15868150 0.79928600
#>  [8,] 1.22051251 0.94466109 0.49214725
#>  [9,] 0.37766102 0.09455561 1.27657776
#> [10,] 1.29236713 0.31783357 1.26002970
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4397588 1.4397588 1.4397588
#>  [2,] 4.5841962 4.5841962 4.5841962
#>  [3,] 1.0794508 1.0794508 1.0794508
#>  [4,] 0.7188771 0.7188771 0.7188771
#>  [5,] 0.2891528 0.2891528 0.2891528
#>  [6,] 3.9731174 3.9731174 3.9731174
#>  [7,] 0.7762676 0.7762676 0.7762676
#>  [8,] 1.7803109 1.7803109 1.7803109
#>  [9,] 1.2399635 1.2399635 1.2399635
#> [10,] 1.2717600 1.2717600 1.2717600

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.19560825 0.13884531 1.39062108
#>  [2,] 2.23130557 0.22212387 0.67222398
#>  [3,] 0.86839019 2.56480183 0.54050003
#>  [4,] 1.29297509 1.77141855 1.77141855
#>  [5,] 0.38850089 0.80553852 0.28108199
#>  [6,] 1.26613603 1.17304244 1.26863901
#>  [7,] 2.16330184 4.23708319 0.91061507
#>  [8,] 1.61157994 1.08329007 1.08329007
#>  [9,] 0.01931377 0.01931377 0.01931377
#> [10,] 0.58963370 0.87862697 0.87862697
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 3.0172358 0.79493749 0.95597035
#>  [2,] 0.5728626 0.02075687 0.01468295
#>  [3,] 4.2610576 1.57574101 0.97851422
#>  [4,] 1.5610638 1.69906928 1.33956476
#>  [5,] 2.6486138 0.13968168 1.37924348
#>  [6,] 0.8007704 1.16832077 0.17335497
#>  [7,] 0.7870742 0.86672085 1.11850062
#>  [8,] 3.2789202 0.64474008 0.95732266
#>  [9,] 0.3403992 3.88159745 1.18191562
#> [10,] 1.2241085 0.11978316 0.38953900
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2089922 0.2089922 0.2089922
#>  [2,] 0.8573963 0.8573963 0.8573963
#>  [3,] 0.3550991 0.3550991 0.3550991
#>  [4,] 0.4425921 0.4425921 0.4425921
#>  [5,] 0.3311717 0.3311717 0.3311717
#>  [6,] 1.1240578 1.1240578 1.1240578
#>  [7,] 3.3196045 3.3196045 3.3196045
#>  [8,] 0.3250940 0.3250940 0.3250940
#>  [9,] 0.8757407 0.8757407 0.8757407
#> [10,] 2.9378948 2.9378948 2.9378948