Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.4479060 0.18000327 0.180003274
#>  [2,] 0.9694024 0.08728460 0.969402414
#>  [3,] 2.9513434 2.49713752 7.078444522
#>  [4,] 4.7842333 0.18189285 1.951922443
#>  [5,] 1.0621602 0.95843613 0.294418309
#>  [6,] 0.6425314 0.81796582 0.006835441
#>  [7,] 0.4148915 0.09166885 0.975299425
#>  [8,] 0.1179902 0.79369005 0.117990173
#>  [9,] 0.1900925 0.19009255 0.190092548
#> [10,] 1.2375334 1.23753343 0.730194764
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.99363589 0.7901999 0.31281100
#>  [2,] 1.40336506 2.2513311 0.05932441
#>  [3,] 1.21035840 1.2725243 0.24803290
#>  [4,] 0.15356319 0.4990025 0.33141409
#>  [5,] 0.84228759 0.4959700 0.22239511
#>  [6,] 0.41543009 2.4574506 0.25620258
#>  [7,] 4.27046844 0.3096383 0.33737198
#>  [8,] 0.62615720 0.2917151 1.35628930
#>  [9,] 0.04409161 0.1375312 0.59882665
#> [10,] 1.98327605 0.4013568 0.81827358
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.1707179 2.1707179 2.1707179
#>  [2,] 0.5716756 0.5716756 0.5716756
#>  [3,] 1.0262793 1.0262793 1.0262793
#>  [4,] 0.2419541 0.2419541 0.2419541
#>  [5,] 3.6323475 3.6323475 3.6323475
#>  [6,] 0.2334741 0.2334741 0.2334741
#>  [7,] 0.1263129 0.1263129 0.1263129
#>  [8,] 0.8957796 0.8957796 0.8957796
#>  [9,] 2.5888281 2.5888281 2.5888281
#> [10,] 1.5614266 1.5614266 1.5614266

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5643965 0.1018869 0.3974537
#>  [2,] 4.5319517 2.6502635 0.5783095
#>  [3,] 1.0799319 1.2169843 1.0799319
#>  [4,] 0.6716731 0.6716731 0.5027608
#>  [5,] 2.1357902 0.4966284 2.1626112
#>  [6,] 0.2611586 0.1047857 1.2169685
#>  [7,] 1.6508401 1.3322954 0.2106569
#>  [8,] 1.8938536 0.7643696 1.8938536
#>  [9,] 2.2841635 2.2841635 0.3891180
#> [10,] 0.4495442 0.4495442 0.4495442
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 1.13879377 0.697734864 0.24667588
#>  [2,] 2.12073301 0.003427527 0.01396087
#>  [3,] 1.33718957 2.765833996 1.52222277
#>  [4,] 0.88670087 0.539123250 0.57802127
#>  [5,] 0.03959048 1.712868650 1.48253420
#>  [6,] 2.19014983 0.815198529 0.99740259
#>  [7,] 0.45089370 0.645023083 0.17505191
#>  [8,] 0.72000069 0.448844001 1.19299080
#>  [9,] 0.25564163 1.485662818 0.34152135
#> [10,] 0.58836686 0.888483368 0.31993976
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 5.09805708 5.09805708 5.09805708
#>  [2,] 3.19760516 3.19760516 3.19760516
#>  [3,] 0.78954402 0.78954402 0.78954402
#>  [4,] 0.19879519 0.19879519 0.19879519
#>  [5,] 0.54279879 0.54279879 0.54279879
#>  [6,] 1.45045248 1.45045248 1.45045248
#>  [7,] 0.43642928 0.43642928 0.43642928
#>  [8,] 0.03919807 0.03919807 0.03919807
#>  [9,] 0.45901437 0.45901437 0.45901437
#> [10,] 0.63123513 0.63123513 0.63123513

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5736977 0.46373222 1.09603528
#>  [2,] 2.2921944 0.76385805 1.12743875
#>  [3,] 1.2009003 2.44141418 1.20090030
#>  [4,] 1.8083802 1.80838021 0.28820918
#>  [5,] 0.9808315 0.98083151 0.07821799
#>  [6,] 1.8335480 0.06165577 0.06165577
#>  [7,] 0.7794466 0.90969051 0.77944664
#>  [8,] 0.6158572 0.11312066 1.27651329
#>  [9,] 0.8462505 0.36671191 0.50940857
#> [10,] 0.3681586 0.98544259 0.36815865
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.65200144 0.169853444 0.19991953
#>  [2,] 0.50237076 3.118462011 1.71355933
#>  [3,] 1.36971173 1.178086366 3.68151266
#>  [4,] 0.03216522 1.563083157 0.49531048
#>  [5,] 4.90740316 0.210826409 1.69785071
#>  [6,] 1.20377069 1.303497889 0.09704478
#>  [7,] 0.22846367 0.333437047 0.97807826
#>  [8,] 0.36312511 0.602522014 0.91938330
#>  [9,] 0.30931840 0.008301798 1.04415554
#> [10,] 2.09443200 0.027970544 4.96870883
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.0282538 0.0282538 0.0282538
#>  [2,] 0.5737950 0.5737950 0.5737950
#>  [3,] 0.5291225 0.5291225 0.5291225
#>  [4,] 4.4440760 4.4440760 4.4440760
#>  [5,] 0.2590821 0.2590821 0.2590821
#>  [6,] 0.2991498 0.2991498 0.2991498
#>  [7,] 1.6009442 1.6009442 1.6009442
#>  [8,] 0.3738054 0.3738054 0.3738054
#>  [9,] 2.0117575 2.0117575 2.0117575
#> [10,] 0.5462514 0.5462514 0.5462514