Sample from extendible Marshall–Olkin distributions

Draws n iid samples from a d-variate extendible Marshall–Olkin distribution parametrized by Bernstein functions bf, essentially wrapping rexmo() by generating suitable exchangeable shock-size-arrival intensities.

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of a extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate extendible Marshall–Olkin distribution with Bernstein function

bf.

Details

The extendible Marshall–Olkin distribution has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} } \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for Bernstein functions \(\psi\), see BernsteinFunction, and \(t_{[1]} \geq \cdots \geq t_{[d]}\), see (Mai and Scherer 2017) . The relationship between Bernstein functions and exchangeable shock-size-arrival intensities of the exchangeable Marshall–Olkin distribution, see rexmo(), is as follows: $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {\{ 1 , \ldots , d \}} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1180811 0.7450991 0.11808113
#>  [2,] 0.2918405 2.8363076 2.83630758
#>  [3,] 0.1510149 0.1510149 0.15101494
#>  [4,] 1.3080365 1.3080365 1.30803653
#>  [5,] 0.2710001 0.2710001 0.03087221
#>  [6,] 0.9690199 0.1298272 0.51451285
#>  [7,] 0.5574306 0.2398066 0.23980656
#>  [8,] 1.6140987 1.1518979 0.45288413
#>  [9,] 1.1289559 0.2230700 0.96562825
#> [10,] 0.2692029 0.2692029 0.26920288
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4112930 1.1290748 2.7162913
#>  [2,] 1.0661975 1.6661624 0.7250253
#>  [3,] 2.6024814 0.4422841 0.8772631
#>  [4,] 0.3229706 1.4324364 0.1927894
#>  [5,] 0.2061017 0.2746498 0.4673253
#>  [6,] 0.4136039 0.9429524 1.7968452
#>  [7,] 0.1399040 0.2013482 0.1064769
#>  [8,] 1.1518230 0.1044213 0.9512335
#>  [9,] 0.3597566 0.2442806 0.5333776
#> [10,] 1.0156244 1.0476572 1.7109449
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1069118 0.1069118 0.1069118
#>  [2,] 1.6691353 1.6691353 1.6691353
#>  [3,] 0.5235189 0.5235189 0.5235189
#>  [4,] 0.6259164 0.6259164 0.6259164
#>  [5,] 1.0799573 1.0799573 1.0799573
#>  [6,] 0.3565611 0.3565611 0.3565611
#>  [7,] 1.8306779 1.8306779 1.8306779
#>  [8,] 1.0695060 1.0695060 1.0695060
#>  [9,] 0.6225094 0.6225094 0.6225094
#> [10,] 0.8635990 0.8635990 0.8635990

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1926973 1.19269733 1.1926973
#>  [2,] 0.2168813 0.53899695 0.2506476
#>  [3,] 0.2309000 0.23090003 0.2309000
#>  [4,] 0.3235701 0.32357012 0.3235701
#>  [5,] 0.1718415 3.45552477 3.4555248
#>  [6,] 0.5685648 0.56856478 0.5685648
#>  [7,] 0.3558372 0.05644975 0.7009312
#>  [8,] 2.2642693 0.98835026 1.6938521
#>  [9,] 0.8154475 0.81544752 1.0453037
#> [10,] 0.8246436 0.82464360 0.8246436
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.01330966 1.8703814 0.7344111
#>  [2,] 1.62733674 0.5463652 3.7950765
#>  [3,] 1.05179529 0.5152421 1.6335491
#>  [4,] 1.63770714 0.7449274 1.2768672
#>  [5,] 0.48945458 1.1141883 0.5534865
#>  [6,] 1.09863480 0.5478402 0.4587924
#>  [7,] 0.56521945 1.3208348 0.4288266
#>  [8,] 1.23511346 0.4908512 0.2952897
#>  [9,] 0.67627604 0.4571328 0.8232235
#> [10,] 0.15277029 2.3214461 1.0749781
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.67700132 0.67700132 0.67700132
#>  [2,] 1.58072053 1.58072053 1.58072053
#>  [3,] 0.39346599 0.39346599 0.39346599
#>  [4,] 0.20401976 0.20401976 0.20401976
#>  [5,] 0.08043203 0.08043203 0.08043203
#>  [6,] 0.14737562 0.14737562 0.14737562
#>  [7,] 0.17694501 0.17694501 0.17694501
#>  [8,] 1.98405392 1.98405392 1.98405392
#>  [9,] 0.97665940 0.97665940 0.97665940
#> [10,] 0.95864549 0.95864549 0.95864549

rextmo(
    10, 3,
    AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40190236 1.68387101 1.68387101
#>  [2,] 0.13896690 0.13896690 0.10504628
#>  [3,] 1.27920695 1.27920695 1.27920695
#>  [4,] 0.69647845 0.69647845 0.69647845
#>  [5,] 0.03153359 0.03153359 0.03153359
#>  [6,] 0.60726209 0.60726209 0.28453605
#>  [7,] 0.40363555 0.40363555 0.58842922
#>  [8,] 0.54782296 2.27589136 0.61217021
#>  [9,] 0.48283718 0.26578513 0.73756894
#> [10,] 1.37729761 1.37729761 1.37729761
# independence
rextmo(
    10, 3,
    LinearBernsteinFunction(scale = 1),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 3.6199396 1.2309511 0.09471612
#>  [2,] 3.6294845 0.5223431 1.17542373
#>  [3,] 0.2968761 1.4845242 1.30884812
#>  [4,] 0.1772773 0.2354139 0.96472072
#>  [5,] 4.1422807 0.9307284 0.11285269
#>  [6,] 0.8337012 3.4899777 1.58015834
#>  [7,] 0.6611224 0.6360343 0.33223563
#>  [8,] 0.4510607 4.3254519 0.69703354
#>  [9,] 0.2244002 1.2079332 0.03629051
#> [10,] 1.3092547 0.8855622 1.24155887
# comonotone
rextmo(
    10, 3,
    ConstantBernsteinFunction(constant = 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.10561857 0.10561857 0.10561857
#>  [2,] 0.72141519 0.72141519 0.72141519
#>  [3,] 1.69302974 1.69302974 1.69302974
#>  [4,] 0.09906027 0.09906027 0.09906027
#>  [5,] 0.91193673 0.91193673 0.91193673
#>  [6,] 0.62661823 0.62661823 0.62661823
#>  [7,] 0.58078652 0.58078652 0.58078652
#>  [8,] 1.22197618 1.22197618 1.22197618
#>  [9,] 0.49924039 0.49924039 0.49924039
#> [10,] 1.55740049 1.55740049 1.55740049