Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

Usage

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0532201 0.1342543 1.0532201
#>  [2,] 2.0148398 0.2441802 2.6586850
#>  [3,] 0.4266984 1.7191085 2.9390995
#>  [4,] 0.9426690 0.9426690 0.4438175
#>  [5,] 0.1646115 0.1646115 0.1646115
#>  [6,] 1.4782288 1.4782288 1.0846330
#>  [7,] 2.2980015 1.1619226 0.2655781
#>  [8,] 3.5831492 2.7489738 2.1674063
#>  [9,] 1.5494220 0.1251068 0.1251068
#> [10,] 2.0019940 2.0019940 1.6700679
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>            [,1]      [,2]      [,3]
#>  [1,] 4.1423828 2.6064893 2.9542080
#>  [2,] 1.0824650 1.1978926 2.0236242
#>  [3,] 2.4289980 2.3791811 1.7018386
#>  [4,] 8.1490808 1.7003975 0.8637372
#>  [5,] 3.8086826 0.5785201 1.3774649
#>  [6,] 1.2515101 1.2236609 0.1229193
#>  [7,] 0.4794320 0.2375766 0.1223151
#>  [8,] 2.9343691 2.8031971 0.1417498
#>  [9,] 1.1016193 0.8391915 0.9554908
#> [10,] 0.1449477 0.8859820 0.1852423
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.07191409 2.07191409 2.07191409
#>  [2,] 2.85272972 2.85272972 2.85272972
#>  [3,] 0.07983915 0.07983915 0.07983915
#>  [4,] 0.95048948 0.95048948 0.95048948
#>  [5,] 0.68816721 0.68816721 0.68816721
#>  [6,] 0.58859513 0.58859513 0.58859513
#>  [7,] 1.37274292 1.37274292 1.37274292
#>  [8,] 0.12326953 0.12326953 0.12326953
#>  [9,] 1.15212072 1.15212072 1.15212072
#> [10,] 0.43963184 0.43963184 0.43963184

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.0489122 0.67767212 1.14013427
#>  [2,] 0.2092223 0.74640140 0.74640140
#>  [3,] 0.8186717 0.40799454 0.40799454
#>  [4,] 0.1919315 0.31871903 0.31871903
#>  [5,] 0.1319511 0.13195112 0.59483982
#>  [6,] 0.1214770 0.03988852 0.03988852
#>  [7,] 4.0091848 0.34358021 0.64328908
#>  [8,] 1.0907235 1.45214794 1.09346964
#>  [9,] 1.0134407 0.08761990 1.01344070
#> [10,] 1.7885214 1.65231790 1.85166308
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]      [,2]        [,3]
#>  [1,] 0.81104924 0.7189879 0.008971597
#>  [2,] 0.79788364 1.6425030 0.848945795
#>  [3,] 1.97440602 3.2176839 0.108775048
#>  [4,] 0.03987006 0.4211554 0.563556523
#>  [5,] 0.98847036 0.3417460 0.387054825
#>  [6,] 1.41946302 1.5226897 0.092915183
#>  [7,] 0.06235749 4.2789361 1.475609705
#>  [8,] 1.12257831 7.0530673 1.090066627
#>  [9,] 2.43882335 0.5031131 0.949605016
#> [10,] 2.45962132 0.1526625 0.210978462
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38142943 0.38142943 0.38142943
#>  [2,] 0.40470628 0.40470628 0.40470628
#>  [3,] 0.24455729 0.24455729 0.24455729
#>  [4,] 2.22033352 2.22033352 2.22033352
#>  [5,] 0.11111658 0.11111658 0.11111658
#>  [6,] 1.13606898 1.13606898 1.13606898
#>  [7,] 0.11233162 0.11233162 0.11233162
#>  [8,] 0.78542817 0.78542817 0.78542817
#>  [9,] 0.01786081 0.01786081 0.01786081
#> [10,] 4.42683958 4.42683958 4.42683958

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1888568 0.1888568 0.1245797
#>  [2,] 1.2796723 0.2243419 1.2796723
#>  [3,] 2.8092597 0.5120205 3.8613233
#>  [4,] 1.2372924 1.2372924 1.2372924
#>  [5,] 3.0041521 2.4297317 2.1171727
#>  [6,] 2.4905037 2.4905037 0.1184049
#>  [7,] 0.5882531 2.9025789 1.6201255
#>  [8,] 3.9786846 3.9786846 3.9786846
#>  [9,] 0.1097987 3.4800915 0.9268327
#> [10,] 0.7104200 0.2065119 1.0749078
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.7400299 0.1654017 1.86780887
#>  [2,] 0.4273744 0.4861880 1.55490141
#>  [3,] 0.5071992 0.7118483 0.31812631
#>  [4,] 0.6920826 0.5293146 1.23300840
#>  [5,] 0.1080440 0.2308830 0.03705070
#>  [6,] 1.3803700 0.2972303 0.12220717
#>  [7,] 0.1847702 1.3211719 2.04375101
#>  [8,] 0.8060620 0.5261800 0.31237274
#>  [9,] 0.5853186 1.1934296 0.09936731
#> [10,] 1.7044325 0.8321938 3.64467120
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.424203436 0.424203436 0.424203436
#>  [2,] 0.002185913 0.002185913 0.002185913
#>  [3,] 1.012895951 1.012895951 1.012895951
#>  [4,] 0.882049224 0.882049224 0.882049224
#>  [5,] 0.976210563 0.976210563 0.976210563
#>  [6,] 0.285615328 0.285615328 0.285615328
#>  [7,] 0.936944500 0.936944500 0.936944500
#>  [8,] 1.604669660 1.604669660 1.604669660
#>  [9,] 1.502499863 1.502499863 1.502499863
#> [10,] 1.156613692 1.156613692 1.156613692