Sample from an extendible Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein function bf by wrapping of rexmo().

rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

bf

a BernsteinFunction with the Bernstein function of the extendible Marshall–Olkin distribution.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-counting model, "AM" for the Arnold model, and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate extendible Marshall–Olkin distribution with Bernstein functionbf.

Parametrization

Extendible Marshall–Olkin distributions are parametrized by Bernstein functions \(\psi\), see (see Sec. 3.3 Mai and Scherer 2017) . The exchangeable shock-size arrival intensities are defined by $$ \eta_{i} = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } , \quad i \in {[d]} . $$

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)))
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0005899 1.0104913 0.06978376
#>  [2,] 1.2609035 0.1299508 0.12995085
#>  [3,] 0.0783950 0.0783950 0.07839500
#>  [4,] 1.4155534 0.2538979 2.21170853
#>  [5,] 0.8704207 1.3426310 1.18510828
#>  [6,] 1.2281641 0.9679610 0.03651702
#>  [7,] 2.1054693 2.1054693 0.10179963
#>  [8,] 1.5749052 1.4895588 1.48955883
#>  [9,] 0.6862308 0.5338022 0.68623080
#> [10,] 3.3575645 0.3254057 3.35756451
rextmo(10, 3, LinearBernsteinFunction(scale = 1))                    # independence
#>             [,1]       [,2]      [,3]
#>  [1,] 4.18544841 1.54033816 1.9520651
#>  [2,] 1.00170519 0.18933867 0.5566175
#>  [3,] 0.06823891 0.96760998 0.6176426
#>  [4,] 0.17339573 0.09025487 0.1138543
#>  [5,] 1.05531461 0.14107400 1.3632254
#>  [6,] 0.22190659 1.60032072 0.3541089
#>  [7,] 0.88279093 1.71020641 0.6867664
#>  [8,] 1.96425870 3.60158467 1.4996549
#>  [9,] 0.34506668 0.96290760 1.9103590
#> [10,] 1.49095216 0.98142783 0.7384750
rextmo(10, 3, ConstantBernsteinFunction(constant = 1))               # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8768265 0.8768265 0.8768265
#>  [2,] 0.4278432 0.4278432 0.4278432
#>  [3,] 0.9927992 0.9927992 0.9927992
#>  [4,] 0.8438687 0.8438687 0.8438687
#>  [5,] 1.9897810 1.9897810 1.9897810
#>  [6,] 0.3361748 0.3361748 0.3361748
#>  [7,] 0.3533650 0.3533650 0.3533650
#>  [8,] 0.5334432 0.5334432 0.5334432
#>  [9,] 0.3668026 0.3668026 0.3668026
#> [10,] 0.2385671 0.2385671 0.2385671

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1195085 1.2113061 1.1603671
#>  [2,] 0.8861479 0.8861479 0.8861479
#>  [3,] 0.6566352 2.6236529 0.6537649
#>  [4,] 0.2650336 0.1076579 3.9918964
#>  [5,] 0.1896514 0.6274142 3.9177269
#>  [6,] 0.3368349 3.8488369 0.3368349
#>  [7,] 0.8210580 0.2257804 0.5173049
#>  [8,] 1.3996846 1.1045679 1.3996846
#>  [9,] 0.4615124 0.4615124 1.0220713
#> [10,] 0.7018199 0.7018199 0.3613773
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "AM")                   # independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.38523828 0.3840453 0.3085851
#>  [2,] 0.47566048 1.2511427 0.4979779
#>  [3,] 1.12095373 0.1286127 1.7449778
#>  [4,] 1.80279169 3.6661623 1.5204768
#>  [5,] 0.68040135 1.1671904 2.5824113
#>  [6,] 0.96728768 0.4525544 0.6346352
#>  [7,] 0.04910092 3.7618793 0.2179156
#>  [8,] 3.51918409 0.5067014 0.3902774
#>  [9,] 0.64935824 0.1836084 0.4934469
#> [10,] 0.02613925 0.6929151 0.9700529
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "AM")              # comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.91668492 0.91668492 0.91668492
#>  [2,] 0.07221914 0.07221914 0.07221914
#>  [3,] 1.06102237 1.06102237 1.06102237
#>  [4,] 0.43850527 0.43850527 0.43850527
#>  [5,] 0.51681460 0.51681460 0.51681460
#>  [6,] 3.28794302 3.28794302 3.28794302
#>  [7,] 0.18411709 0.18411709 0.18411709
#>  [8,] 0.79605562 0.79605562 0.79605562
#>  [9,] 0.79367049 0.79367049 0.79367049
#> [10,] 1.80515754 1.80515754 1.80515754

rextmo(10, 3, AlphaStableBernsteinFunction(alpha = log2(2  - 0.5)), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4124862 0.4124862 0.3256597
#>  [2,] 0.3730067 1.2480293 1.2480293
#>  [3,] 1.7360582 0.7061583 0.8498197
#>  [4,] 2.0244260 2.0244260 2.1484237
#>  [5,] 0.1779207 0.3158326 2.3643114
#>  [6,] 2.0178747 1.0495001 1.0495001
#>  [7,] 0.7354116 0.7354116 0.7354116
#>  [8,] 2.1718523 0.4962662 4.6086532
#>  [9,] 0.4374962 0.3345194 0.3345194
#> [10,] 0.9483709 0.2016590 0.2016590
rextmo(10, 3, LinearBernsteinFunction(scale = 1), method = "ESM")                  # independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2468246 0.73167575 0.1158698
#>  [2,] 0.3478380 0.46556663 0.2240946
#>  [3,] 1.0447076 1.00520445 3.2036828
#>  [4,] 0.3116533 0.01762835 0.4841629
#>  [5,] 1.7010078 0.34672571 0.6494618
#>  [6,] 1.6911571 0.08604141 0.6882147
#>  [7,] 1.8148965 2.24622043 0.1007017
#>  [8,] 2.5921108 2.86527621 1.0488845
#>  [9,] 1.9757249 0.09915788 0.7202694
#> [10,] 2.5782498 0.12387281 1.2160546
rextmo(10, 3, ConstantBernsteinFunction(constant = 1), method = "ESM")             # comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5679601 1.5679601 1.5679601
#>  [2,] 0.1610740 0.1610740 0.1610740
#>  [3,] 0.3932162 0.3932162 0.3932162
#>  [4,] 1.7805869 1.7805869 1.7805869
#>  [5,] 0.2864080 0.2864080 0.2864080
#>  [6,] 0.4895917 0.4895917 0.4895917
#>  [7,] 3.8965219 3.8965219 3.8965219
#>  [8,] 0.7534406 0.7534406 0.7534406
#>  [9,] 1.1376633 1.1376633 1.1376633
#> [10,] 0.2736132 0.2736132 0.2736132