Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.25324065 3.24810253 0.52879675
#>  [2,] 0.79421169 0.23007106 0.79421169
#>  [3,] 0.08411387 0.16038327 0.08600139
#>  [4,] 0.48164695 0.48164695 0.48164695
#>  [5,] 0.21914890 0.73292751 0.73292751
#>  [6,] 1.51412453 1.08231870 2.79775807
#>  [7,] 3.25202627 0.81189533 0.35590507
#>  [8,] 0.23734455 0.06998938 0.20539648
#>  [9,] 0.14508990 2.70030818 2.39923085
#> [10,] 0.14085764 0.14085764 0.14085764
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.91137514 0.56603118 0.95056568
#>  [2,] 3.65710185 1.36918994 0.09364676
#>  [3,] 1.64794508 0.08456794 2.67202898
#>  [4,] 0.50842925 1.19485904 8.33411254
#>  [5,] 0.00769766 0.15820328 0.98152558
#>  [6,] 1.17033852 0.46797866 0.06065143
#>  [7,] 0.18823307 2.93493887 1.72267227
#>  [8,] 0.65155970 0.61680130 0.10883879
#>  [9,] 0.13839047 0.79312686 0.58572747
#> [10,] 0.83126850 1.69202928 1.17646327
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9425833 1.9425833 1.9425833
#>  [2,] 0.0594114 0.0594114 0.0594114
#>  [3,] 0.7874468 0.7874468 0.7874468
#>  [4,] 0.2711525 0.2711525 0.2711525
#>  [5,] 0.0139113 0.0139113 0.0139113
#>  [6,] 1.7175581 1.7175581 1.7175581
#>  [7,] 0.2632245 0.2632245 0.2632245
#>  [8,] 2.1408503 2.1408503 2.1408503
#>  [9,] 0.3749771 0.3749771 0.3749771
#> [10,] 0.3308688 0.3308688 0.3308688

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 2.232074846 2.71430630 0.129904842
#>  [2,] 0.156928812 0.15692881 0.156928812
#>  [3,] 0.009871246 1.48123572 0.009871246
#>  [4,] 3.021001015 3.02100101 6.381351352
#>  [5,] 1.285870827 1.23277603 1.061148593
#>  [6,] 0.609650295 4.91023249 0.481989968
#>  [7,] 1.225090098 1.22509010 0.851464911
#>  [8,] 0.546814757 1.33214368 0.546814757
#>  [9,] 0.319692455 1.65847031 0.341545347
#> [10,] 0.078808723 0.07880872 0.078808723
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6577076 1.5131668 0.09705446
#>  [2,] 2.7266348 1.1111670 1.56077511
#>  [3,] 0.3723933 0.9132184 4.65089666
#>  [4,] 1.6264767 0.8721914 0.94721091
#>  [5,] 2.9332468 0.2210538 1.08395633
#>  [6,] 0.7230848 0.2317664 1.02718372
#>  [7,] 1.0491141 0.2462353 0.22290711
#>  [8,] 0.9230302 0.1465861 2.86904163
#>  [9,] 4.3292245 1.7667274 0.32330069
#> [10,] 1.0591875 0.4533012 0.43505317
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2984105 0.2984105 0.2984105
#>  [2,] 0.3772686 0.3772686 0.3772686
#>  [3,] 0.2765848 0.2765848 0.2765848
#>  [4,] 0.4192172 0.4192172 0.4192172
#>  [5,] 4.7646995 4.7646995 4.7646995
#>  [6,] 2.3304185 2.3304185 2.3304185
#>  [7,] 1.5681882 1.5681882 1.5681882
#>  [8,] 0.4976759 0.4976759 0.4976759
#>  [9,] 1.3331459 1.3331459 1.3331459
#> [10,] 0.2849041 0.2849041 0.2849041

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1135342 0.07899145 0.07899145
#>  [2,] 1.4849309 1.48493088 1.48493088
#>  [3,] 0.8873698 0.42044640 0.68142297
#>  [4,] 1.3222998 0.07019809 2.52661955
#>  [5,] 0.6449569 0.64495685 0.64495685
#>  [6,] 0.1403777 0.45616072 0.51533904
#>  [7,] 0.3647567 0.36475674 0.36475674
#>  [8,] 0.3214831 0.32148311 0.32148311
#>  [9,] 1.0615635 3.16170925 0.29060626
#> [10,] 1.3193331 1.31933306 1.31933306
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6773771 0.82504726 1.1256632
#>  [2,] 0.2342922 0.58679161 1.4154284
#>  [3,] 0.3030079 4.20476387 0.1890063
#>  [4,] 2.8412997 0.07945408 3.7082035
#>  [5,] 0.1207108 0.13029431 0.4965017
#>  [6,] 1.6068982 0.26636986 0.6630642
#>  [7,] 1.0205717 4.07256854 1.6862142
#>  [8,] 2.5641239 3.35428617 0.6479818
#>  [9,] 0.5633307 0.19477918 0.5631463
#> [10,] 0.8290851 0.49332635 0.6677970
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.76970574 0.76970574 0.76970574
#>  [2,] 0.97078341 0.97078341 0.97078341
#>  [3,] 0.17569099 0.17569099 0.17569099
#>  [4,] 0.14205815 0.14205815 0.14205815
#>  [5,] 0.11351367 0.11351367 0.11351367
#>  [6,] 0.04566964 0.04566964 0.04566964
#>  [7,] 0.11917659 0.11917659 0.11917659
#>  [8,] 6.61568592 6.61568592 6.61568592
#>  [9,] 4.66064383 4.66064383 4.66064383
#> [10,] 1.84624199 1.84624199 1.84624199