Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.59996593 1.04465477 0.01686493
#>  [2,] 0.41997063 0.57020104 0.41997063
#>  [3,] 0.03119118 0.03119118 0.03119118
#>  [4,] 0.66491882 0.66491882 0.78596233
#>  [5,] 0.44911723 1.62953428 2.78154883
#>  [6,] 0.83790148 1.47056355 0.83790148
#>  [7,] 1.71036271 1.55170855 2.81750473
#>  [8,] 0.44347350 0.11570118 3.25371661
#>  [9,] 0.86422633 0.54589496 0.17883821
#> [10,] 1.26031910 0.14027889 1.14481200
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.57467801 0.4638028 0.7751668
#>  [2,] 0.09385575 1.2629188 1.1644179
#>  [3,] 0.45433815 0.2108391 2.6103651
#>  [4,] 2.14730413 0.5886844 0.5781608
#>  [5,] 2.30774145 0.3750704 0.7891073
#>  [6,] 0.76702155 0.3504294 0.3152249
#>  [7,] 1.77574719 0.5555528 0.6155655
#>  [8,] 1.25934686 0.5573505 1.2866633
#>  [9,] 0.61933683 2.5708453 4.4833358
#> [10,] 1.41700268 0.8203620 4.0557160
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9975968 0.9975968 0.9975968
#>  [2,] 0.3847524 0.3847524 0.3847524
#>  [3,] 0.1059570 0.1059570 0.1059570
#>  [4,] 0.1196454 0.1196454 0.1196454
#>  [5,] 2.0465890 2.0465890 2.0465890
#>  [6,] 0.1694189 0.1694189 0.1694189
#>  [7,] 0.6000506 0.6000506 0.6000506
#>  [8,] 0.1845600 0.1845600 0.1845600
#>  [9,] 0.7292393 0.7292393 0.7292393
#> [10,] 0.8676089 0.8676089 0.8676089

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04402011 0.02927301 0.04402011
#>  [2,] 0.91357711 0.93683583 0.31306185
#>  [3,] 0.49720106 0.22924317 0.49720106
#>  [4,] 0.03572685 0.03572685 0.03572685
#>  [5,] 0.13010938 0.13010938 0.46477290
#>  [6,] 0.51947913 0.23988644 0.91172951
#>  [7,] 0.53807644 0.53807644 0.53807644
#>  [8,] 0.31087000 0.31087000 0.31087000
#>  [9,] 0.80401895 0.03605893 0.03605893
#> [10,] 0.24920210 0.24920210 0.24920210
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 2.35203432 1.0157618 0.92252758
#>  [2,] 0.51710841 1.5982086 2.88869644
#>  [3,] 2.10830699 1.0930041 0.78462667
#>  [4,] 2.00835467 0.8581021 3.13630784
#>  [5,] 4.26372081 0.1334031 0.37471026
#>  [6,] 1.91564953 0.5471883 0.20826806
#>  [7,] 0.29152035 2.3015092 0.02624406
#>  [8,] 0.86994192 0.6721276 1.30569528
#>  [9,] 1.66930096 0.0629233 2.66949513
#> [10,] 0.03798393 0.7679092 2.21605804
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.38091882 2.38091882 2.38091882
#>  [2,] 1.66982293 1.66982293 1.66982293
#>  [3,] 0.47489813 0.47489813 0.47489813
#>  [4,] 0.93607562 0.93607562 0.93607562
#>  [5,] 0.14175240 0.14175240 0.14175240
#>  [6,] 0.80817800 0.80817800 0.80817800
#>  [7,] 0.05403060 0.05403060 0.05403060
#>  [8,] 1.31941713 1.31941713 1.31941713
#>  [9,] 0.07295702 0.07295702 0.07295702
#> [10,] 0.27277005 0.27277005 0.27277005

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]        [,2]        [,3]
#>  [1,] 0.5437835 0.339417201 0.543783451
#>  [2,] 0.3093237 0.931583832 1.032456258
#>  [3,] 0.1598964 0.001667875 0.001667875
#>  [4,] 0.5074045 0.507404536 0.507404536
#>  [5,] 0.5284352 0.528435240 2.780682363
#>  [6,] 3.0284425 4.553112402 0.173187897
#>  [7,] 0.4069867 0.212027007 0.518964352
#>  [8,] 2.1369153 3.484820593 0.014595426
#>  [9,] 0.8119044 0.241845639 0.909010822
#> [10,] 0.4666317 0.466631696 0.466631696
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>              [,1]      [,2]       [,3]
#>  [1,] 1.351197590 1.2429851 3.04155070
#>  [2,] 0.161822142 0.6965295 3.36769095
#>  [3,] 0.062821133 0.2765965 0.65820062
#>  [4,] 0.184671847 0.3674328 0.55301169
#>  [5,] 2.816438836 0.1025502 0.11110003
#>  [6,] 0.237762045 0.1030205 0.05943189
#>  [7,] 0.568807444 1.5642401 1.20117973
#>  [8,] 0.274849630 0.7514725 1.10888504
#>  [9,] 0.007407186 0.6084909 0.64824705
#> [10,] 0.393391268 0.2480907 1.45038737
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7478461 0.7478461 0.7478461
#>  [2,] 0.1766570 0.1766570 0.1766570
#>  [3,] 1.1213883 1.1213883 1.1213883
#>  [4,] 0.1415212 0.1415212 0.1415212
#>  [5,] 0.1975498 0.1975498 0.1975498
#>  [6,] 0.3808096 0.3808096 0.3808096
#>  [7,] 2.2391437 2.2391437 2.2391437
#>  [8,] 0.3283852 0.3283852 0.3283852
#>  [9,] 0.2004854 0.2004854 0.2004854
#> [10,] 0.1109326 0.1109326 0.1109326