Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04060876 0.04060876 0.04060876
#>  [2,] 0.90444147 0.90444147 0.90444147
#>  [3,] 1.12401365 0.71467566 0.46196585
#>  [4,] 1.27175649 0.99370400 1.27175649
#>  [5,] 1.98200607 1.98200607 1.68324640
#>  [6,] 1.72080709 1.72080709 0.21420485
#>  [7,] 2.89368752 3.76129822 0.82859390
#>  [8,] 3.99493969 4.32633540 5.39533792
#>  [9,] 0.01093037 0.86459793 0.01093037
#> [10,] 1.30378548 0.75108861 4.89385563
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7061183 0.83970857 1.5003022
#>  [2,] 2.6222083 4.04085238 0.3129242
#>  [3,] 0.1411128 1.79022403 1.0218980
#>  [4,] 2.2021875 0.10992702 0.8335702
#>  [5,] 0.4293376 0.93281263 0.4176732
#>  [6,] 1.5550628 0.06771931 0.9225040
#>  [7,] 2.4041425 0.95798903 1.2999489
#>  [8,] 1.5519102 0.58910688 2.5157536
#>  [9,] 0.6799320 1.34076653 0.3558846
#> [10,] 0.4389832 0.19190723 0.5517872
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1805310 1.1805310 1.1805310
#>  [2,] 2.3070002 2.3070002 2.3070002
#>  [3,] 0.2232250 0.2232250 0.2232250
#>  [4,] 0.4058456 0.4058456 0.4058456
#>  [5,] 1.3573795 1.3573795 1.3573795
#>  [6,] 0.2870996 0.2870996 0.2870996
#>  [7,] 2.3788248 2.3788248 2.3788248
#>  [8,] 0.5807747 0.5807747 0.5807747
#>  [9,] 0.3046914 0.3046914 0.3046914
#> [10,] 0.9459614 0.9459614 0.9459614

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1444382 1.3727879 2.6813781
#>  [2,] 1.5924193 0.9351447 1.5924193
#>  [3,] 0.3221217 0.3221217 0.3221217
#>  [4,] 0.6999021 0.4585195 0.4585195
#>  [5,] 0.7444185 1.5747919 1.5747919
#>  [6,] 0.6553470 1.9277914 0.3536443
#>  [7,] 0.2498203 0.5329177 2.7201284
#>  [8,] 0.6000929 0.5331256 0.6000929
#>  [9,] 0.2764946 2.6792168 1.3290979
#> [10,] 0.9683274 0.9683274 0.9683274
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 3.7453695 0.73055843 0.77780311
#>  [2,] 1.1877318 1.60036236 0.13720911
#>  [3,] 0.9846201 0.35069602 1.10399932
#>  [4,] 0.9944133 0.89585452 0.46587082
#>  [5,] 1.5632270 0.36319311 1.84240166
#>  [6,] 0.3899881 0.07054733 2.56353863
#>  [7,] 0.4383038 0.33208775 0.55472132
#>  [8,] 1.8100816 1.88040942 0.40591316
#>  [9,] 0.8499135 0.33229322 0.78302096
#> [10,] 0.1864707 0.80823867 0.06933432
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6028185 1.6028185 1.6028185
#>  [2,] 1.0017915 1.0017915 1.0017915
#>  [3,] 0.1498282 0.1498282 0.1498282
#>  [4,] 1.0540733 1.0540733 1.0540733
#>  [5,] 1.1883413 1.1883413 1.1883413
#>  [6,] 0.6221417 0.6221417 0.6221417
#>  [7,] 0.5512987 0.5512987 0.5512987
#>  [8,] 0.1676039 0.1676039 0.1676039
#>  [9,] 1.7377367 1.7377367 1.7377367
#> [10,] 1.0251123 1.0251123 1.0251123

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.56526632 0.1893314 0.61921390
#>  [2,] 2.15879404 0.3961014 0.31532463
#>  [3,] 0.45128060 0.7977268 0.01893592
#>  [4,] 1.59747746 1.5974775 1.59747746
#>  [5,] 0.40083176 0.4008318 0.40083176
#>  [6,] 0.06548026 1.9571523 1.02111693
#>  [7,] 0.98780651 1.6629528 1.66295277
#>  [8,] 4.31765532 2.6329401 1.68619209
#>  [9,] 0.29001990 0.6045627 0.61594270
#> [10,] 0.85113760 0.8511376 0.85113760
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2187228 0.08162214 0.23466357
#>  [2,] 0.4158974 3.53374322 0.43123440
#>  [3,] 0.8459453 2.05205024 0.48246708
#>  [4,] 0.6140785 2.30068612 0.65134444
#>  [5,] 1.1656002 0.38348735 2.11786464
#>  [6,] 0.8989155 0.34957171 0.92282500
#>  [7,] 2.2708254 4.99109746 0.44807166
#>  [8,] 2.0283976 0.51034814 0.36052193
#>  [9,] 0.3241329 2.35664929 0.99040423
#> [10,] 1.4003746 0.07277712 0.02349723
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09069197 0.09069197 0.09069197
#>  [2,] 0.13013636 0.13013636 0.13013636
#>  [3,] 1.27314089 1.27314089 1.27314089
#>  [4,] 1.29550937 1.29550937 1.29550937
#>  [5,] 0.73955256 0.73955256 0.73955256
#>  [6,] 0.54591535 0.54591535 0.54591535
#>  [7,] 0.54806226 0.54806226 0.54806226
#>  [8,] 0.17109556 0.17109556 0.17109556
#>  [9,] 0.46936648 0.46936648 0.46936648
#> [10,] 0.02164982 0.02164982 0.02164982