Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]      [,2]      [,3]
#>  [1,] 1.30957678 1.3095768 1.3095768
#>  [2,] 1.46340926 4.0658754 0.7445027
#>  [3,] 1.10016721 0.6480425 1.1001672
#>  [4,] 0.50257478 2.0461790 0.7449921
#>  [5,] 0.04645795 0.7017207 0.4987518
#>  [6,] 0.45201427 0.1775085 1.8546706
#>  [7,] 1.48930978 2.9052832 2.9052832
#>  [8,] 1.48253847 4.6580671 1.3633263
#>  [9,] 0.39838650 1.1881225 1.1881225
#> [10,] 0.10600563 0.1060056 0.1060056
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5888517 0.28009846 0.69914406
#>  [2,] 1.6817749 1.06054024 0.33937058
#>  [3,] 0.3020425 0.06604321 2.66472492
#>  [4,] 1.0038044 0.14743382 0.02848314
#>  [5,] 0.3122080 1.07933004 1.61309036
#>  [6,] 1.1410124 0.04154307 0.26407207
#>  [7,] 0.4162964 0.43976998 0.41787015
#>  [8,] 1.1310885 0.84088676 1.61437179
#>  [9,] 1.5133904 0.25615115 2.21714720
#> [10,] 1.6461269 0.46165215 0.21543251
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.67236685 0.67236685 0.67236685
#>  [2,] 0.09638279 0.09638279 0.09638279
#>  [3,] 1.03395640 1.03395640 1.03395640
#>  [4,] 2.71621766 2.71621766 2.71621766
#>  [5,] 0.29679846 0.29679846 0.29679846
#>  [6,] 0.31776360 0.31776360 0.31776360
#>  [7,] 0.43847691 0.43847691 0.43847691
#>  [8,] 0.22485535 0.22485535 0.22485535
#>  [9,] 1.36858688 1.36858688 1.36858688
#> [10,] 1.82651475 1.82651475 1.82651475

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6231062 0.32945872 0.3294587
#>  [2,] 1.0117600 1.81753795 6.3069336
#>  [3,] 0.9613791 1.25039853 1.5742582
#>  [4,] 3.1928093 4.16979669 0.2981122
#>  [5,] 1.3401278 1.16094633 1.1609463
#>  [6,] 0.4311916 0.43119156 0.4311916
#>  [7,] 0.4345127 0.24756149 0.5257591
#>  [8,] 0.7350024 0.73500244 0.9042560
#>  [9,] 1.4117083 0.01637565 1.4117083
#> [10,] 3.7016894 3.39137067 1.1290060
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>             [,1]         [,2]       [,3]
#>  [1,] 3.75327257 0.0151020698 1.53885774
#>  [2,] 0.74967663 2.0679723359 0.58182554
#>  [3,] 0.08391260 1.7274508573 2.66272190
#>  [4,] 1.08284851 0.0905967556 0.72290241
#>  [5,] 0.02541719 0.0439416777 1.17459933
#>  [6,] 0.28050597 0.2320569721 3.07158971
#>  [7,] 1.40940223 0.8751990394 0.06343974
#>  [8,] 0.28377672 0.0004022848 0.53646313
#>  [9,] 0.69249408 0.3832391296 0.16458144
#> [10,] 1.62993294 0.1073981474 1.14785466
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.954392327 0.954392327 0.954392327
#>  [2,] 0.050521363 0.050521363 0.050521363
#>  [3,] 3.500031452 3.500031452 3.500031452
#>  [4,] 0.764185034 0.764185034 0.764185034
#>  [5,] 0.715927440 0.715927440 0.715927440
#>  [6,] 0.813681293 0.813681293 0.813681293
#>  [7,] 1.690890627 1.690890627 1.690890627
#>  [8,] 0.050569912 0.050569912 0.050569912
#>  [9,] 0.933323433 0.933323433 0.933323433
#> [10,] 0.001676961 0.001676961 0.001676961

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.35827060 1.39298070 1.2254616
#>  [2,] 0.03182465 0.03182465 0.1235569
#>  [3,] 3.66612330 0.97652417 0.2448614
#>  [4,] 1.81575593 1.81575593 1.8157559
#>  [5,] 0.07673910 0.07673910 0.0767391
#>  [6,] 4.05407454 0.96614268 2.3663731
#>  [7,] 2.33294848 3.11134832 0.4439611
#>  [8,] 0.73927232 0.73927232 0.7392723
#>  [9,] 0.22868518 2.75370367 2.3956590
#> [10,] 0.66609565 0.17652978 0.6660956
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4480913 1.4027549 0.2376519
#>  [2,] 0.1572566 0.5017196 2.1113474
#>  [3,] 0.1020940 0.1807527 1.7043546
#>  [4,] 2.9907080 1.1388465 0.9907354
#>  [5,] 0.2431304 0.9878309 0.1595049
#>  [6,] 0.9357476 1.5374690 2.4299609
#>  [7,] 0.8805000 2.9358440 1.2322632
#>  [8,] 0.7693970 0.5589825 0.8223816
#>  [9,] 0.1173710 0.4371837 0.4826656
#> [10,] 2.3979071 0.6674924 1.9160544
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7650347 0.7650347 0.7650347
#>  [2,] 0.1813189 0.1813189 0.1813189
#>  [3,] 1.0412039 1.0412039 1.0412039
#>  [4,] 0.5574705 0.5574705 0.5574705
#>  [5,] 1.8052325 1.8052325 1.8052325
#>  [6,] 1.8439856 1.8439856 1.8439856
#>  [7,] 0.1701183 0.1701183 0.1701183
#>  [8,] 0.6710153 0.6710153 0.6710153
#>  [9,] 0.6324555 0.6324555 0.6324555
#> [10,] 2.1724820 2.1724820 2.1724820