Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.08168317 1.08168317 1.08168317
#>  [2,] 0.45927996 0.45927996 0.45927996
#>  [3,] 0.18778883 0.18778883 0.18778883
#>  [4,] 0.70921790 1.61998743 0.13782660
#>  [5,] 0.54656602 0.20520970 2.32862739
#>  [6,] 0.14228916 1.23448813 0.10822263
#>  [7,] 2.32431003 2.32431003 0.99189661
#>  [8,] 0.05810556 0.05810556 0.05810556
#>  [9,] 1.31187058 1.46208470 1.46208470
#> [10,] 0.67646664 0.64654454 0.35249777
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.02170188 0.02153389 1.4438989
#>  [2,] 0.15426352 0.50016562 1.3849443
#>  [3,] 1.65491432 0.40026982 1.6924650
#>  [4,] 1.90943105 0.71912732 0.9106896
#>  [5,] 1.24065940 0.14780057 0.3853049
#>  [6,] 0.48700035 0.74672656 1.9139485
#>  [7,] 1.63523515 0.02133264 5.1754609
#>  [8,] 2.73178197 1.29150144 0.3175104
#>  [9,] 2.08489780 0.88594788 1.4940531
#> [10,] 3.19920473 0.88277493 0.5107442
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.443042487 0.443042487 0.443042487
#>  [2,] 0.146344983 0.146344983 0.146344983
#>  [3,] 1.167478664 1.167478664 1.167478664
#>  [4,] 0.006733552 0.006733552 0.006733552
#>  [5,] 0.803324780 0.803324780 0.803324780
#>  [6,] 0.761245985 0.761245985 0.761245985
#>  [7,] 0.329245966 0.329245966 0.329245966
#>  [8,] 0.356730125 0.356730125 0.356730125
#>  [9,] 1.623622088 1.623622088 1.623622088
#> [10,] 5.730335816 5.730335816 5.730335816

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2671633 0.25886849 0.2671633
#>  [2,] 0.5853201 0.58532006 0.5853201
#>  [3,] 3.8810346 3.88103457 1.5993023
#>  [4,] 1.5005032 0.96363078 0.6912284
#>  [5,] 1.2328040 0.56732112 1.2328040
#>  [6,] 0.1424330 0.05351995 0.1424330
#>  [7,] 1.4526799 3.76439256 1.0819804
#>  [8,] 0.9139759 0.23063654 0.9139759
#>  [9,] 0.7625777 0.56381835 1.1155009
#> [10,] 0.4456348 0.19491728 0.4456348
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.54253016 0.33171797 0.6133424
#>  [2,] 0.73875303 0.02935234 4.0406653
#>  [3,] 0.56588771 0.39808488 2.1455288
#>  [4,] 0.11301240 1.99333448 2.1060638
#>  [5,] 0.09264802 0.83235046 2.8747121
#>  [6,] 1.12563423 0.38401702 0.2399338
#>  [7,] 1.22680733 1.18302692 0.8643142
#>  [8,] 0.00374809 3.22498263 1.3555537
#>  [9,] 2.11614159 0.77178152 0.9584490
#> [10,] 0.46267304 0.18990740 1.4055781
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3263865 1.3263865 1.3263865
#>  [2,] 5.7641790 5.7641790 5.7641790
#>  [3,] 0.6563622 0.6563622 0.6563622
#>  [4,] 3.0146774 3.0146774 3.0146774
#>  [5,] 1.6343551 1.6343551 1.6343551
#>  [6,] 0.6067526 0.6067526 0.6067526
#>  [7,] 0.8214051 0.8214051 0.8214051
#>  [8,] 0.6963983 0.6963983 0.6963983
#>  [9,] 0.1456790 0.1456790 0.1456790
#> [10,] 2.5581527 2.5581527 2.5581527

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8024790 0.8024790 0.23682678
#>  [2,] 0.2477783 0.2477783 0.24777833
#>  [3,] 2.3455336 0.9054437 3.01743206
#>  [4,] 0.4097849 0.4097849 0.08273332
#>  [5,] 0.4418511 0.4418511 0.15966290
#>  [6,] 4.8821830 0.1420371 0.14203706
#>  [7,] 1.0797315 0.5016877 0.09140239
#>  [8,] 0.6346748 0.8363877 0.83638770
#>  [9,] 0.1750697 0.1750697 0.17506966
#> [10,] 1.7412468 0.2974122 1.74124683
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1510365 1.0035864 0.6251762
#>  [2,] 1.7039546 1.5230331 0.1341254
#>  [3,] 1.6588390 0.4410140 0.2337592
#>  [4,] 1.3761951 0.1358952 1.7201255
#>  [5,] 0.6138325 0.3390501 2.9499819
#>  [6,] 3.2509168 1.2252417 1.2747140
#>  [7,] 2.1980514 3.0638447 0.2725128
#>  [8,] 0.3246716 0.6278127 0.1134340
#>  [9,] 5.1987081 0.9396175 2.0737588
#> [10,] 0.4806002 0.2406036 1.3177479
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.43005772 2.43005772 2.43005772
#>  [2,] 0.36107644 0.36107644 0.36107644
#>  [3,] 1.63483582 1.63483582 1.63483582
#>  [4,] 0.53528751 0.53528751 0.53528751
#>  [5,] 0.02378753 0.02378753 0.02378753
#>  [6,] 1.23776828 1.23776828 1.23776828
#>  [7,] 1.01672509 1.01672509 1.01672509
#>  [8,] 0.26638342 0.26638342 0.26638342
#>  [9,] 0.40524358 0.40524358 0.40524358
#> [10,] 0.11428827 0.11428827 0.11428827