Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.07545321 0.01922278 1.1156633
#>  [2,] 2.03266620 0.70798302 0.9559626
#>  [3,] 0.77872036 1.50252057 2.0024351
#>  [4,] 0.28497832 0.28497832 0.2849783
#>  [5,] 0.30061228 0.30061228 2.7862288
#>  [6,] 1.22022655 0.44270518 0.3738823
#>  [7,] 1.70223399 0.77783742 1.7022340
#>  [8,] 0.33035946 0.33035946 1.1176813
#>  [9,] 0.03406521 0.03406521 0.3160816
#> [10,] 0.63483795 0.16832241 3.0136594
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.07325247 0.1772360 2.10121448
#>  [2,] 2.31433472 0.5834957 0.70613754
#>  [3,] 0.17959131 0.3631227 0.05942139
#>  [4,] 0.16261343 0.2341107 0.04468613
#>  [5,] 3.42239466 0.6161738 0.66518040
#>  [6,] 0.91687172 0.4796792 0.10385476
#>  [7,] 0.47479962 0.2265739 1.39229027
#>  [8,] 0.03019445 1.5530662 0.06679356
#>  [9,] 0.91635243 4.2036713 2.03838438
#> [10,] 0.33447533 0.4291142 0.87445740
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4649509 0.4649509 0.4649509
#>  [2,] 1.9806879 1.9806879 1.9806879
#>  [3,] 0.2749859 0.2749859 0.2749859
#>  [4,] 1.0103052 1.0103052 1.0103052
#>  [5,] 0.7546813 0.7546813 0.7546813
#>  [6,] 0.1152009 0.1152009 0.1152009
#>  [7,] 0.3942552 0.3942552 0.3942552
#>  [8,] 0.4539071 0.4539071 0.4539071
#>  [9,] 2.0182584 2.0182584 2.0182584
#> [10,] 2.2051069 2.2051069 2.2051069

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.38309160 0.3830916 0.38309160
#>  [2,] 1.77757723 0.3773433 0.05182073
#>  [3,] 0.06972874 3.2939085 2.28571100
#>  [4,] 0.47280372 0.2186781 0.84238400
#>  [5,] 2.57102879 1.6284663 2.57102879
#>  [6,] 1.78377020 2.1722552 0.58344436
#>  [7,] 1.54101760 0.2935707 1.50540267
#>  [8,] 1.03388001 0.7671759 0.76717590
#>  [9,] 0.43141763 0.4314176 0.92051813
#> [10,] 0.53302401 0.5330240 0.53302401
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2044327 1.0077043 0.09377687
#>  [2,] 0.6181971 0.6719598 0.78519582
#>  [3,] 0.1537048 0.1085777 0.60039550
#>  [4,] 3.9044141 0.9633613 1.87018785
#>  [5,] 0.5790118 1.0770782 0.94483601
#>  [6,] 0.9211137 0.6088563 0.13456953
#>  [7,] 0.1565319 0.5325400 1.49808420
#>  [8,] 0.5956122 0.4280694 0.53641288
#>  [9,] 3.4197501 0.6378929 0.16951580
#> [10,] 0.1473376 0.5505411 0.37219842
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4009407 0.4009407 0.4009407
#>  [2,] 0.4356961 0.4356961 0.4356961
#>  [3,] 1.1272005 1.1272005 1.1272005
#>  [4,] 0.7994613 0.7994613 0.7994613
#>  [5,] 0.9214727 0.9214727 0.9214727
#>  [6,] 0.3583572 0.3583572 0.3583572
#>  [7,] 0.1280032 0.1280032 0.1280032
#>  [8,] 0.5199811 0.5199811 0.5199811
#>  [9,] 1.3491378 1.3491378 1.3491378
#> [10,] 1.0199359 1.0199359 1.0199359

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9055056 0.9055056 0.9055056
#>  [2,] 0.8269180 0.3350424 0.3350424
#>  [3,] 1.7028187 1.4340147 1.3276784
#>  [4,] 2.0189669 1.0933006 2.3147627
#>  [5,] 0.1914359 0.3425285 0.1545319
#>  [6,] 0.3975496 0.3975496 0.4785794
#>  [7,] 0.2486425 0.2486425 0.2486425
#>  [8,] 2.8483175 0.8430224 0.9951708
#>  [9,] 1.0649867 1.4671293 1.0848592
#> [10,] 0.7553050 2.6821810 0.3738985
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5406463 3.03358721 2.04891872
#>  [2,] 0.8712057 1.04827269 0.24215700
#>  [3,] 0.5805594 2.15979792 1.31309789
#>  [4,] 3.2004896 0.96878251 0.15733534
#>  [5,] 1.8330388 0.61684662 0.62393739
#>  [6,] 1.4781383 0.45638112 1.40136884
#>  [7,] 0.7054514 0.47807705 3.57796080
#>  [8,] 0.9064224 0.02551085 0.07988148
#>  [9,] 0.8086691 1.27259004 1.43711904
#> [10,] 0.9239903 1.19097745 1.81607595
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41333663 0.41333663 0.41333663
#>  [2,] 1.72829730 1.72829730 1.72829730
#>  [3,] 0.20528401 0.20528401 0.20528401
#>  [4,] 2.09901071 2.09901071 2.09901071
#>  [5,] 1.16544131 1.16544131 1.16544131
#>  [6,] 0.86858937 0.86858937 0.86858937
#>  [7,] 1.15215508 1.15215508 1.15215508
#>  [8,] 2.03673961 2.03673961 2.03673961
#>  [9,] 1.11400244 1.11400244 1.11400244
#> [10,] 0.01057058 0.01057058 0.01057058