Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>               [,1]         [,2]         [,3]
#>  [1,] 0.7462744055 0.6538876306 0.2487707241
#>  [2,] 0.0007202849 0.0007202849 0.0007202849
#>  [3,] 1.2390602993 1.2390602993 0.0483436174
#>  [4,] 2.1781874781 0.3077213287 1.3711815707
#>  [5,] 0.5560772680 0.5560772680 0.5560772680
#>  [6,] 0.6026560970 0.7313154071 0.7313154071
#>  [7,] 0.1926976945 0.0778471459 0.1926976945
#>  [8,] 0.0273967240 0.0273967240 0.0273967240
#>  [9,] 0.9769709261 0.7666100071 0.9769709261
#> [10,] 0.7057678338 0.2642060593 0.6542030228
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.78949055 1.57543617 0.1855978
#>  [2,] 0.55190614 1.36144719 0.1735493
#>  [3,] 0.48792083 1.32278583 0.1214395
#>  [4,] 1.13009760 2.96573648 0.5046500
#>  [5,] 0.07564629 3.30365783 0.2725335
#>  [6,] 1.48745251 1.92261027 1.0084979
#>  [7,] 0.27060882 0.77156593 0.6396069
#>  [8,] 1.87211447 0.69087654 0.7921210
#>  [9,] 2.15145404 0.04407503 1.8123092
#> [10,] 0.92167820 4.07121518 0.1160113
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.89593662 0.89593662 0.89593662
#>  [2,] 2.03138547 2.03138547 2.03138547
#>  [3,] 0.48374469 0.48374469 0.48374469
#>  [4,] 0.13119837 0.13119837 0.13119837
#>  [5,] 1.05384081 1.05384081 1.05384081
#>  [6,] 1.85044312 1.85044312 1.85044312
#>  [7,] 0.44875480 0.44875480 0.44875480
#>  [8,] 0.19273572 0.19273572 0.19273572
#>  [9,] 0.09486099 0.09486099 0.09486099
#> [10,] 1.92253780 1.92253780 1.92253780

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.10413067 0.78515494 0.78515494
#>  [2,] 2.63683589 2.13022820 2.29491984
#>  [3,] 0.81566891 2.99784422 0.81566891
#>  [4,] 1.90121189 1.68446716 1.94125446
#>  [5,] 1.95436667 3.65847724 0.21656165
#>  [6,] 0.09077777 0.22792517 0.22792517
#>  [7,] 1.25307181 1.07875129 1.07875129
#>  [8,] 1.95197101 1.95197101 1.95197101
#>  [9,] 1.42092682 0.06199959 0.06199959
#> [10,] 1.83070386 4.15977077 1.83070386
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.0966244 0.04064243 0.16956114
#>  [2,] 1.4942131 1.44681993 1.61977886
#>  [3,] 2.1856552 0.27913852 1.43129840
#>  [4,] 0.1859048 0.31027294 0.19137462
#>  [5,] 1.5150700 1.21581574 0.06497078
#>  [6,] 2.7353460 1.29734907 0.32189607
#>  [7,] 0.3404355 0.30685181 0.72780732
#>  [8,] 0.4976680 0.47244066 3.33850113
#>  [9,] 0.4437573 2.70713473 0.06778842
#> [10,] 0.6689885 0.02950921 0.33304404
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6031620 0.6031620 0.6031620
#>  [2,] 0.2230730 0.2230730 0.2230730
#>  [3,] 1.1783406 1.1783406 1.1783406
#>  [4,] 0.7347799 0.7347799 0.7347799
#>  [5,] 0.4020061 0.4020061 0.4020061
#>  [6,] 3.2987486 3.2987486 3.2987486
#>  [7,] 1.1127582 1.1127582 1.1127582
#>  [8,] 0.4855560 0.4855560 0.4855560
#>  [9,] 2.0682982 2.0682982 2.0682982
#> [10,] 1.4278880 1.4278880 1.4278880

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2872935 1.2872935 1.5757186
#>  [2,] 1.1403573 1.0938056 1.1403573
#>  [3,] 0.3897622 1.3256157 3.1311697
#>  [4,] 0.4102103 0.4102103 0.4102103
#>  [5,] 1.3960113 0.8928070 1.3960113
#>  [6,] 0.8756623 0.4064613 0.8756623
#>  [7,] 0.9287905 0.8798958 4.9428172
#>  [8,] 0.9869404 0.1054547 1.6976739
#>  [9,] 1.4091771 0.9865158 1.4091771
#> [10,] 2.4129862 2.4129862 1.9393504
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.4353275 0.02625074 4.01134256
#>  [2,] 1.3546008 0.94741398 0.15429333
#>  [3,] 0.6675316 1.53936930 0.42284538
#>  [4,] 0.1145819 0.31056091 1.71870542
#>  [5,] 0.6609434 1.80141256 1.28047689
#>  [6,] 0.5079368 0.81527928 0.34028310
#>  [7,] 1.1194546 0.76778479 0.09685601
#>  [8,] 0.9044043 1.47975588 0.40465751
#>  [9,] 1.0173708 0.95950018 0.31844368
#> [10,] 2.2724675 0.36660605 1.32831282
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8689825 0.8689825 0.8689825
#>  [2,] 0.3774774 0.3774774 0.3774774
#>  [3,] 0.8301367 0.8301367 0.8301367
#>  [4,] 2.7882905 2.7882905 2.7882905
#>  [5,] 0.1130081 0.1130081 0.1130081
#>  [6,] 0.2676675 0.2676675 0.2676675
#>  [7,] 0.1880310 0.1880310 0.1880310
#>  [8,] 0.6673620 0.6673620 0.6673620
#>  [9,] 0.1511120 0.1511120 0.1511120
#> [10,] 0.7930649 0.7930649 0.7930649