Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]      [,2]      [,3]
#>  [1,] 0.17696938 0.1769694 0.1769694
#>  [2,] 0.66179865 0.6617986 0.6617986
#>  [3,] 0.12642027 0.1264203 0.1264203
#>  [4,] 0.94509688 0.9450969 0.9450969
#>  [5,] 0.18606840 0.2932801 0.2932801
#>  [6,] 0.04517212 0.3764606 0.3764606
#>  [7,] 0.90758910 0.9075891 0.9075891
#>  [8,] 1.54597169 1.9432184 1.4059352
#>  [9,] 0.79299646 1.7994409 2.4101054
#> [10,] 0.44902826 1.1272932 0.4490283
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.89109476 0.4085878 2.96608977
#>  [2,] 0.50255512 2.4399238 0.64846596
#>  [3,] 1.04234321 2.3364375 0.17194618
#>  [4,] 0.07720035 0.8194980 0.92823015
#>  [5,] 2.55602080 1.6965300 1.19476040
#>  [6,] 1.44167551 2.2188750 1.57340512
#>  [7,] 0.75297734 1.8723686 0.05097561
#>  [8,] 2.93183272 0.5446078 0.49364911
#>  [9,] 0.49290428 0.6455910 0.31652273
#> [10,] 0.65695634 0.2882532 0.79424755
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.88261319 0.88261319 0.88261319
#>  [2,] 0.43127549 0.43127549 0.43127549
#>  [3,] 3.13332320 3.13332320 3.13332320
#>  [4,] 1.91170546 1.91170546 1.91170546
#>  [5,] 1.31819152 1.31819152 1.31819152
#>  [6,] 2.24102025 2.24102025 2.24102025
#>  [7,] 2.03719120 2.03719120 2.03719120
#>  [8,] 0.04233659 0.04233659 0.04233659
#>  [9,] 0.61828627 0.61828627 0.61828627
#> [10,] 1.28932258 1.28932258 1.28932258

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.27768052 0.28808627 2.01466352
#>  [2,] 1.35636714 0.17009042 0.17009042
#>  [3,] 1.63238470 1.77884056 1.77884056
#>  [4,] 0.20714989 0.20714989 0.20714989
#>  [5,] 1.32088431 1.32088431 0.33484047
#>  [6,] 0.96720424 0.92333679 0.66344219
#>  [7,] 0.25007304 0.47236233 0.82188114
#>  [8,] 1.44655803 0.87009550 1.29280912
#>  [9,] 0.11328334 0.11328334 4.11720795
#> [10,] 0.02231527 0.02231527 0.02231527
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.60539953 0.28958690 0.0730063
#>  [2,] 0.17775169 1.53937161 0.5498691
#>  [3,] 0.05153333 0.13148462 0.5500037
#>  [4,] 0.63397962 3.88084037 0.1652353
#>  [5,] 1.99497686 0.03538249 0.0484038
#>  [6,] 0.73749833 0.56782027 0.3359440
#>  [7,] 0.46870572 2.16101509 0.7440124
#>  [8,] 1.11978225 1.70474448 0.1786986
#>  [9,] 0.69148950 0.02988648 0.9713462
#> [10,] 0.92047978 0.93699750 0.1788931
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 5.6078790 5.6078790 5.6078790
#>  [2,] 3.3316229 3.3316229 3.3316229
#>  [3,] 0.4059910 0.4059910 0.4059910
#>  [4,] 0.3962277 0.3962277 0.3962277
#>  [5,] 1.4925347 1.4925347 1.4925347
#>  [6,] 1.6904764 1.6904764 1.6904764
#>  [7,] 1.5041932 1.5041932 1.5041932
#>  [8,] 0.8444204 0.8444204 0.8444204
#>  [9,] 1.6507021 1.6507021 1.6507021
#> [10,] 3.7010927 3.7010927 3.7010927

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18346710 1.35683303 1.52149520
#>  [2,] 0.41863304 0.26446298 0.26446298
#>  [3,] 1.05665263 1.05665263 1.05665263
#>  [4,] 0.01084441 0.01084441 0.01084441
#>  [5,] 0.52632446 0.60214089 0.60214089
#>  [6,] 0.11895704 2.33297402 0.11895704
#>  [7,] 0.33815697 0.33815697 0.33815697
#>  [8,] 0.73096991 2.33568207 0.73096991
#>  [9,] 0.75345625 3.35363923 3.35363923
#> [10,] 2.44321227 1.60627616 2.44321227
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3130571 0.6105698 0.9750140
#>  [2,] 1.0887306 1.1146244 3.0174483
#>  [3,] 0.2135654 0.1309048 1.0510030
#>  [4,] 1.0306662 0.6623466 2.2258186
#>  [5,] 1.1382374 0.2222471 0.8751834
#>  [6,] 0.5000031 1.0734304 1.6228185
#>  [7,] 0.9289832 3.5303794 0.1400265
#>  [8,] 0.6927903 0.4330352 0.6933330
#>  [9,] 1.0126547 1.5864131 0.1058271
#> [10,] 0.5000180 2.5963268 1.1288842
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3959399 0.3959399 0.3959399
#>  [2,] 0.8999050 0.8999050 0.8999050
#>  [3,] 0.9688939 0.9688939 0.9688939
#>  [4,] 0.5462020 0.5462020 0.5462020
#>  [5,] 0.1472547 0.1472547 0.1472547
#>  [6,] 0.5218736 0.5218736 0.5218736
#>  [7,] 0.4752589 0.4752589 0.4752589
#>  [8,] 0.3046490 0.3046490 0.3046490
#>  [9,] 0.2327527 0.2327527 0.2327527
#> [10,] 3.3204042 3.3204042 3.3204042