Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0943309 0.2600333 0.6060134
#>  [2,] 0.8117601 0.8117601 2.2236336
#>  [3,] 2.6700419 0.4451948 0.4451948
#>  [4,] 0.8988907 0.3638461 1.6853106
#>  [5,] 0.8265733 0.8265733 0.1834316
#>  [6,] 2.1922375 0.1148920 1.0016078
#>  [7,] 1.6842300 1.8197375 1.8197375
#>  [8,] 0.2145711 0.2145711 0.2145711
#>  [9,] 1.3485475 1.3485475 0.5726317
#> [10,] 0.2221357 0.2221357 1.2192781
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>             [,1]        [,2]      [,3]
#>  [1,] 0.55374091 0.005471506 2.8762529
#>  [2,] 1.09436251 0.237101367 0.3079095
#>  [3,] 0.40902610 0.090976721 3.0622290
#>  [4,] 1.03818666 0.010766964 0.3433316
#>  [5,] 3.88981301 1.867773408 0.1209302
#>  [6,] 0.40873572 0.316492818 0.6805366
#>  [7,] 0.05628397 0.717708084 0.9585148
#>  [8,] 0.51130754 3.440312375 4.1811100
#>  [9,] 0.36067995 2.307650794 0.3751122
#> [10,] 0.23781029 0.025676659 0.5237555
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14529955 0.14529955 0.14529955
#>  [2,] 0.23504863 0.23504863 0.23504863
#>  [3,] 0.41002738 0.41002738 0.41002738
#>  [4,] 0.33786285 0.33786285 0.33786285
#>  [5,] 0.03507782 0.03507782 0.03507782
#>  [6,] 0.06311369 0.06311369 0.06311369
#>  [7,] 1.82817267 1.82817267 1.82817267
#>  [8,] 0.01592931 0.01592931 0.01592931
#>  [9,] 1.44125628 1.44125628 1.44125628
#> [10,] 2.02499126 2.02499126 2.02499126

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.04302999 1.04302999 1.04302999
#>  [2,] 0.08961649 0.08961649 0.08961649
#>  [3,] 0.85308255 0.85308255 0.45856868
#>  [4,] 1.22894257 1.22894257 1.22894257
#>  [5,] 0.96928953 0.96928953 0.96928953
#>  [6,] 1.17846465 0.30303653 2.92831843
#>  [7,] 0.25281034 0.25281034 0.25281034
#>  [8,] 0.53701410 0.26454727 0.53701410
#>  [9,] 2.82272536 3.46762331 0.50433571
#> [10,] 0.30918251 0.30918251 0.30918251
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2418181 0.20932540 0.5019532
#>  [2,] 0.3020719 0.56215891 2.2373099
#>  [3,] 0.1544708 0.01138795 0.8250215
#>  [4,] 1.2140427 1.72809332 0.2245727
#>  [5,] 1.9645348 0.91846428 2.3713718
#>  [6,] 2.7567966 2.28118686 0.1781978
#>  [7,] 1.7964654 2.79409487 0.2136740
#>  [8,] 0.8184090 1.36192472 0.6270253
#>  [9,] 0.2182799 1.54808177 0.1105426
#> [10,] 0.3156608 0.84189689 1.4620630
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55758664 0.55758664 0.55758664
#>  [2,] 0.90357430 0.90357430 0.90357430
#>  [3,] 1.62874024 1.62874024 1.62874024
#>  [4,] 0.70920609 0.70920609 0.70920609
#>  [5,] 1.59619611 1.59619611 1.59619611
#>  [6,] 0.83756103 0.83756103 0.83756103
#>  [7,] 0.06220571 0.06220571 0.06220571
#>  [8,] 1.48740447 1.48740447 1.48740447
#>  [9,] 0.82892831 0.82892831 0.82892831
#> [10,] 0.43701047 0.43701047 0.43701047

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.63136801 1.38016680 2.27660399
#>  [2,] 3.42454380 0.54515447 3.88721735
#>  [3,] 0.28365773 0.28365773 0.28365773
#>  [4,] 3.06112110 3.74340369 3.74340369
#>  [5,] 0.08444295 0.08444295 0.08444295
#>  [6,] 0.07248388 0.07248388 0.07248388
#>  [7,] 0.40579087 0.76308220 0.76308220
#>  [8,] 0.41916487 1.41218474 1.41218474
#>  [9,] 0.31814313 0.31814313 0.31814313
#> [10,] 0.26430863 0.26430863 0.26430863
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.38840978 0.4420527 0.83795642
#>  [2,] 0.04219132 0.4352749 1.01423299
#>  [3,] 0.39791974 2.0230835 0.57890121
#>  [4,] 0.45331590 0.7851116 1.19129885
#>  [5,] 1.80587574 0.9259720 0.02309979
#>  [6,] 1.35758227 0.5098781 0.21860858
#>  [7,] 0.53386032 0.9118414 1.37698161
#>  [8,] 0.70648729 0.1216292 0.12918283
#>  [9,] 0.49178886 0.2816361 0.31337492
#> [10,] 0.93775703 2.8362992 0.07420818
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5356680 1.5356680 1.5356680
#>  [2,] 0.2181274 0.2181274 0.2181274
#>  [3,] 0.1850057 0.1850057 0.1850057
#>  [4,] 1.3552947 1.3552947 1.3552947
#>  [5,] 0.2164910 0.2164910 0.2164910
#>  [6,] 0.1585477 0.1585477 0.1585477
#>  [7,] 0.4758343 0.4758343 0.4758343
#>  [8,] 1.5707610 1.5707610 1.5707610
#>  [9,] 0.1666013 0.1666013 0.1666013
#> [10,] 1.3997355 1.3997355 1.3997355