Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM for small dimensions only; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns, with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.69794486 0.69794486 0.69794486
#>  [2,] 0.09117840 0.09117840 0.09117840
#>  [3,] 0.30607839 5.41857329 0.30607839
#>  [4,] 0.09657520 0.09657520 0.09657520
#>  [5,] 2.34336810 0.67961732 3.17802264
#>  [6,] 0.19442735 0.19442735 0.19442735
#>  [7,] 2.31570703 2.31570703 2.31570703
#>  [8,] 0.03234796 1.21751160 0.03234796
#>  [9,] 0.52400185 0.52400185 0.52400185
#> [10,] 0.04545061 0.04545061 0.04545061
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.75267655 1.5846262 0.64390298
#>  [2,] 0.17622997 0.6608812 0.01277703
#>  [3,] 0.12538524 0.1070958 3.23471227
#>  [4,] 1.56879155 1.6496390 0.40057848
#>  [5,] 0.05993640 1.1195105 3.71428535
#>  [6,] 1.05351387 2.2829132 0.89420512
#>  [7,] 0.05788872 0.9503894 0.09329146
#>  [8,] 0.45036453 1.2162857 1.37634707
#>  [9,] 1.06543798 0.1596119 0.96542858
#> [10,] 1.29455151 0.8918193 0.61854326
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.921713516 1.921713516 1.921713516
#>  [2,] 0.790373957 0.790373957 0.790373957
#>  [3,] 0.005633799 0.005633799 0.005633799
#>  [4,] 0.665533995 0.665533995 0.665533995
#>  [5,] 0.587846890 0.587846890 0.587846890
#>  [6,] 0.117956505 0.117956505 0.117956505
#>  [7,] 3.176403612 3.176403612 3.176403612
#>  [8,] 0.511486922 0.511486922 0.511486922
#>  [9,] 0.545474425 0.545474425 0.545474425
#> [10,] 2.347528439 2.347528439 2.347528439

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58986999 0.92010951 0.58986999
#>  [2,] 1.44242759 0.70019168 0.70019168
#>  [3,] 1.47296405 1.52468902 1.22523023
#>  [4,] 0.09473871 0.93714000 0.30708493
#>  [5,] 0.25216826 0.03643048 0.25216826
#>  [6,] 0.62884099 0.80044475 0.80044475
#>  [7,] 1.74131785 1.88283760 2.09235364
#>  [8,] 0.89445318 0.00898341 0.89445318
#>  [9,] 2.62070490 0.47360961 0.45554367
#> [10,] 1.06606895 0.26962558 0.04558519
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02380529 1.51266091 0.61252622
#>  [2,] 0.20772595 0.85084451 1.14562106
#>  [3,] 0.54830056 0.20551582 2.12018922
#>  [4,] 0.60688937 0.07131045 0.17588497
#>  [5,] 0.71903919 0.78350702 0.90749629
#>  [6,] 0.69542735 0.34153151 3.23436225
#>  [7,] 3.97472181 0.63155406 0.09924942
#>  [8,] 0.07810458 4.26373355 1.96757842
#>  [9,] 0.89360748 0.24680023 0.43966187
#> [10,] 2.15615705 0.54282015 1.06068594
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55278530 0.55278530 0.55278530
#>  [2,] 0.10372151 0.10372151 0.10372151
#>  [3,] 0.19696644 0.19696644 0.19696644
#>  [4,] 0.16423315 0.16423315 0.16423315
#>  [5,] 0.05734189 0.05734189 0.05734189
#>  [6,] 4.42198098 4.42198098 4.42198098
#>  [7,] 0.98940095 0.98940095 0.98940095
#>  [8,] 0.48289217 0.48289217 0.48289217
#>  [9,] 2.43690808 2.43690808 2.43690808
#> [10,] 0.50326570 0.50326570 0.50326570

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36533719 0.53626804 0.36533719
#>  [2,] 0.61062626 0.61062626 0.61062626
#>  [3,] 2.28847596 4.22674758 2.63668619
#>  [4,] 0.56562071 1.28030887 1.28030887
#>  [5,] 1.08315173 1.76277994 1.08315173
#>  [6,] 2.07366993 0.39934811 1.77374837
#>  [7,] 0.39509377 0.39509377 0.39509377
#>  [8,] 0.05196488 0.05196488 0.05196488
#>  [9,] 2.80462192 1.96634217 0.34027212
#> [10,] 1.37171042 1.74997202 1.50848016
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>             [,1]      [,2]         [,3]
#>  [1,] 0.06936003 1.2240047 2.5422902333
#>  [2,] 0.32752593 1.9776755 0.2923765290
#>  [3,] 0.12818446 1.0374629 1.6027431725
#>  [4,] 1.87985598 0.8404653 0.7840233935
#>  [5,] 1.12844407 0.5898079 0.0008534871
#>  [6,] 1.18196362 0.7666070 0.6490627688
#>  [7,] 0.41866089 0.3080343 1.4979418723
#>  [8,] 0.17986474 1.5467218 0.9700144999
#>  [9,] 0.35736140 0.3337628 2.8405129145
#> [10,] 1.60268079 0.2645971 1.1878052280
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06383001 0.06383001 0.06383001
#>  [2,] 0.76957787 0.76957787 0.76957787
#>  [3,] 2.57873203 2.57873203 2.57873203
#>  [4,] 0.52308856 0.52308856 0.52308856
#>  [5,] 0.82452530 0.82452530 0.82452530
#>  [6,] 1.39913994 1.39913994 1.39913994
#>  [7,] 0.11771737 0.11771737 0.11771737
#>  [8,] 0.02131036 0.02131036 0.02131036
#>  [9,] 1.14989388 1.14989388 1.14989388
#> [10,] 0.04721746 0.04721746 0.04721746