Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47996580 1.01251949 1.26178113
#>  [2,] 0.65096278 0.65096278 0.65096278
#>  [3,] 1.81498036 0.41965362 1.81498036
#>  [4,] 0.75226215 3.33835204 0.75226215
#>  [5,] 0.82307541 0.42817351 0.48524591
#>  [6,] 0.55025898 0.71967257 0.73210144
#>  [7,] 0.60716665 1.06306179 0.01604651
#>  [8,] 0.17200715 0.17200715 0.17200715
#>  [9,] 0.26522626 0.54657012 0.87750365
#> [10,] 0.01088046 0.01088046 0.01088046
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>            [,1]        [,2]      [,3]
#>  [1,] 0.2459566 0.369641486 0.4361961
#>  [2,] 0.5811277 0.304943131 0.5064175
#>  [3,] 0.1199376 0.185611448 2.5277492
#>  [4,] 1.1544820 1.052777629 0.2913479
#>  [5,] 2.3648827 0.675363987 0.2185655
#>  [6,] 2.1398954 1.508848518 0.1072698
#>  [7,] 0.2429071 1.119007858 0.1187548
#>  [8,] 1.4004812 0.004909699 0.2376164
#>  [9,] 1.4826237 1.001034525 0.1836507
#> [10,] 0.3202706 1.839763312 0.9646800
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 2.06746383 2.06746383 2.06746383
#>  [2,] 1.49614242 1.49614242 1.49614242
#>  [3,] 0.06535763 0.06535763 0.06535763
#>  [4,] 1.02690778 1.02690778 1.02690778
#>  [5,] 0.77510028 0.77510028 0.77510028
#>  [6,] 2.02391796 2.02391796 2.02391796
#>  [7,] 1.01390325 1.01390325 1.01390325
#>  [8,] 0.62717374 0.62717374 0.62717374
#>  [9,] 0.41749133 0.41749133 0.41749133
#> [10,] 2.33702351 2.33702351 2.33702351

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41304911 0.41304911 0.41304911
#>  [2,] 0.96889032 0.96889032 1.53893620
#>  [3,] 0.73384085 0.96874305 0.40931494
#>  [4,] 0.31813182 0.31813182 0.44224887
#>  [5,] 0.33672288 0.33672288 0.33672288
#>  [6,] 0.16925373 0.16925373 0.16925373
#>  [7,] 1.09325060 2.00225546 0.59719549
#>  [8,] 4.56898805 1.98606996 2.37977102
#>  [9,] 0.02201326 0.02201326 0.02201326
#> [10,] 0.08250365 0.08250365 0.08250365
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.72389768 0.15684902 1.8263830
#>  [2,] 0.06302033 0.37189980 2.2336417
#>  [3,] 2.14246143 0.04410848 0.3455992
#>  [4,] 0.14449610 0.35242326 0.2068297
#>  [5,] 0.02423391 0.69721742 3.4256264
#>  [6,] 2.33036497 0.57479647 1.7840222
#>  [7,] 0.33437332 1.97623924 0.3175252
#>  [8,] 0.22992481 0.78264246 2.5994688
#>  [9,] 0.19760962 1.10118517 0.4031670
#> [10,] 1.77480755 4.08978840 0.8945215
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 5.14222448 5.14222448 5.14222448
#>  [2,] 0.54634430 0.54634430 0.54634430
#>  [3,] 0.17430507 0.17430507 0.17430507
#>  [4,] 0.92094637 0.92094637 0.92094637
#>  [5,] 0.18354437 0.18354437 0.18354437
#>  [6,] 0.01926841 0.01926841 0.01926841
#>  [7,] 1.85368347 1.85368347 1.85368347
#>  [8,] 0.09847206 0.09847206 0.09847206
#>  [9,] 0.15264221 0.15264221 0.15264221
#> [10,] 0.89622425 0.89622425 0.89622425

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.99848514 0.89869270 0.76193782
#>  [2,] 2.70763972 2.70763972 0.24136099
#>  [3,] 0.96371143 0.96371143 0.85154294
#>  [4,] 1.46917914 1.46917914 1.46917914
#>  [5,] 0.04097822 0.04097822 0.04097822
#>  [6,] 0.36383294 1.84333150 0.36383294
#>  [7,] 1.69084932 0.18953313 0.18953313
#>  [8,] 0.30218929 0.66985601 1.01333824
#>  [9,] 0.73427201 0.73427201 0.73427201
#> [10,] 0.26344955 0.26344955 0.26344955
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 2.67217227 1.68326095 1.2718904
#>  [2,] 0.37140335 0.31128447 0.2372430
#>  [3,] 2.71371904 0.08728020 0.7479380
#>  [4,] 0.41112228 0.30788846 0.9495204
#>  [5,] 0.23856530 0.22178527 0.1844478
#>  [6,] 0.33650393 2.55508628 1.6623080
#>  [7,] 2.73155934 2.06293924 2.1809891
#>  [8,] 1.32185508 0.67916336 2.1063485
#>  [9,] 1.40024701 0.07321433 1.1645619
#> [10,] 0.03442896 0.02727953 1.3488033
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1589987 0.1589987 0.1589987
#>  [2,] 0.1415684 0.1415684 0.1415684
#>  [3,] 1.1546531 1.1546531 1.1546531
#>  [4,] 1.8217982 1.8217982 1.8217982
#>  [5,] 1.9071703 1.9071703 1.9071703
#>  [6,] 0.2114840 0.2114840 0.2114840
#>  [7,] 1.2825313 1.2825313 1.2825313
#>  [8,] 0.3752004 0.3752004 0.3752004
#>  [9,] 1.1482790 1.1482790 1.1482790
#> [10,] 0.1145999 0.1145999 0.1145999