Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.50181813 1.12671670 1.1267167
#>  [2,] 1.90757616 3.49316935 0.4820691
#>  [3,] 0.27284446 0.26476277 1.2702955
#>  [4,] 1.57263102 0.24127394 1.1351873
#>  [5,] 2.40470660 2.40470660 0.2225437
#>  [6,] 1.31523980 0.40146379 3.1358691
#>  [7,] 0.47607670 0.47607670 0.4760767
#>  [8,] 2.06324741 0.36981490 0.9549772
#>  [9,] 1.76107813 0.40684387 0.3417760
#> [10,] 0.04893323 0.04893323 1.5244872
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]      [,2]        [,3]
#>  [1,] 0.5567310 0.8392818 0.006716962
#>  [2,] 0.7227359 0.2264880 0.335060511
#>  [3,] 2.1204958 0.2833484 0.918859664
#>  [4,] 0.5753148 0.6481626 0.145099602
#>  [5,] 1.1901123 0.3863603 1.643321088
#>  [6,] 0.3073978 0.3084248 2.141651051
#>  [7,] 0.2159958 1.6847576 1.700539357
#>  [8,] 0.3780209 2.7029530 0.788322225
#>  [9,] 0.5587022 1.0428836 0.413824521
#> [10,] 0.3591463 0.7872284 0.889305984
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7998565 0.7998565 0.7998565
#>  [2,] 0.9288087 0.9288087 0.9288087
#>  [3,] 0.2786563 0.2786563 0.2786563
#>  [4,] 0.2010312 0.2010312 0.2010312
#>  [5,] 1.3351009 1.3351009 1.3351009
#>  [6,] 0.6673205 0.6673205 0.6673205
#>  [7,] 0.3000429 0.3000429 0.3000429
#>  [8,] 0.5863441 0.5863441 0.5863441
#>  [9,] 0.1104645 0.1104645 0.1104645
#> [10,] 1.0573372 1.0573372 1.0573372

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 1.31567456 0.880599906 0.70053093
#>  [2,] 0.07480107 0.075470014 0.07547001
#>  [3,] 2.07752950 1.695011812 1.69501181
#>  [4,] 0.96089024 1.442876587 0.96089024
#>  [5,] 2.84598342 0.627846094 1.43657265
#>  [6,] 1.02320235 0.091163018 0.12015092
#>  [7,] 1.37151602 2.303373634 1.87203443
#>  [8,] 1.18145020 1.181450195 0.86524927
#>  [9,] 1.00181736 0.003911173 4.35303064
#> [10,] 2.70603562 2.617359201 2.10423547
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4556306 0.3734465 1.4640599
#>  [2,] 0.5406911 2.2882899 0.6703615
#>  [3,] 0.3227550 3.9225416 1.0936953
#>  [4,] 2.0483984 0.4398980 0.2546078
#>  [5,] 1.7162881 0.2328856 0.8605668
#>  [6,] 0.6786044 3.2985064 0.2704079
#>  [7,] 1.4047640 2.5544629 1.3034899
#>  [8,] 1.8244396 2.0291555 0.2035869
#>  [9,] 0.6370002 0.5209425 0.8778290
#> [10,] 0.4283828 0.6825598 1.3039563
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2087041 0.2087041 0.2087041
#>  [2,] 0.5283857 0.5283857 0.5283857
#>  [3,] 0.5073594 0.5073594 0.5073594
#>  [4,] 0.1679477 0.1679477 0.1679477
#>  [5,] 3.1693280 3.1693280 3.1693280
#>  [6,] 2.8704308 2.8704308 2.8704308
#>  [7,] 1.9056899 1.9056899 1.9056899
#>  [8,] 0.4131612 0.4131612 0.4131612
#>  [9,] 1.3807687 1.3807687 1.3807687
#> [10,] 0.7478443 0.7478443 0.7478443

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.61517249 3.78419674 0.55500660
#>  [2,] 0.33666080 0.28810252 0.84501234
#>  [3,] 0.09459181 0.09459181 0.51048509
#>  [4,] 4.16428443 2.11780599 2.17381253
#>  [5,] 0.88820867 0.96934559 0.63494025
#>  [6,] 0.25792102 0.25792102 0.25792102
#>  [7,] 0.50428024 0.60983562 0.74088666
#>  [8,] 0.04662642 0.33493465 0.04662642
#>  [9,] 0.16383590 0.16383590 0.16383590
#> [10,] 1.00142876 1.00142876 1.00142876
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3170203 0.59782822 0.6935970
#>  [2,] 1.7743504 0.10165052 1.2787154
#>  [3,] 3.4331834 3.23930959 0.7287025
#>  [4,] 0.2817761 0.92169831 0.4251526
#>  [5,] 3.5067753 1.13165799 2.5807441
#>  [6,] 0.3185142 0.76387502 1.3510559
#>  [7,] 0.3481233 0.09943206 0.2525095
#>  [8,] 1.0895698 0.32639820 5.1217994
#>  [9,] 0.8660430 0.67632682 1.3252028
#> [10,] 0.1780140 2.10734044 0.3253184
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6612508 1.6612508 1.6612508
#>  [2,] 1.1834558 1.1834558 1.1834558
#>  [3,] 0.1008928 0.1008928 0.1008928
#>  [4,] 0.6441948 0.6441948 0.6441948
#>  [5,] 1.0208507 1.0208507 1.0208507
#>  [6,] 0.4324211 0.4324211 0.4324211
#>  [7,] 1.1038384 1.1038384 1.1038384
#>  [8,] 1.0182343 1.0182343 1.0182343
#>  [9,] 2.1341906 2.1341906 2.1341906
#> [10,] 0.1805122 0.1805122 0.1805122