Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>            [,1]        [,2]      [,3]
#>  [1,] 0.2992610 7.353342750 0.9550751
#>  [2,] 0.8873712 0.002441735 1.1949178
#>  [3,] 0.6208330 0.620832951 0.6208330
#>  [4,] 0.7843474 0.597446178 0.7843474
#>  [5,] 1.4564168 0.982604715 0.7432808
#>  [6,] 0.2614849 0.261484881 0.2614849
#>  [7,] 0.2436803 0.243680320 1.3486957
#>  [8,] 1.4086337 0.374082732 0.4912659
#>  [9,] 0.3853537 0.385353703 0.3853537
#> [10,] 1.0082740 0.507893333 0.3903004
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>             [,1]      [,2]      [,3]
#>  [1,] 0.02564616 1.1746990 1.8436365
#>  [2,] 0.15090669 0.3922790 0.2466309
#>  [3,] 0.17386075 0.6980861 1.3773164
#>  [4,] 0.20354149 0.8323733 0.7463430
#>  [5,] 0.50542700 2.0546513 2.3747092
#>  [6,] 2.66630800 3.0306846 0.1888618
#>  [7,] 1.01207397 0.1688781 4.8128433
#>  [8,] 3.75414624 3.1728247 0.4169165
#>  [9,] 0.93615609 1.0478378 1.0890170
#> [10,] 0.97290237 1.2988721 0.4069375
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5091967 1.5091967 1.5091967
#>  [2,] 0.8064633 0.8064633 0.8064633
#>  [3,] 9.2042337 9.2042337 9.2042337
#>  [4,] 3.1685660 3.1685660 3.1685660
#>  [5,] 1.2372902 1.2372902 1.2372902
#>  [6,] 0.2788293 0.2788293 0.2788293
#>  [7,] 2.1711054 2.1711054 2.1711054
#>  [8,] 2.4336107 2.4336107 2.4336107
#>  [9,] 1.0531985 1.0531985 1.0531985
#> [10,] 4.7929955 4.7929955 4.7929955

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.575951486 0.575951486 0.253305116
#>  [2,] 0.322716860 1.291709258 0.322716860
#>  [3,] 1.173683099 0.660114367 2.951826099
#>  [4,] 2.257935327 4.169899225 2.855256312
#>  [5,] 0.198353664 0.198353664 0.198353664
#>  [6,] 0.454385772 2.374616216 2.374616216
#>  [7,] 0.007775358 0.007775358 0.007775358
#>  [8,] 0.276472243 0.276472243 0.276472243
#>  [9,] 1.239238167 0.406358605 0.406358605
#> [10,] 1.779158485 1.715044451 3.286596106
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.006166262 0.03828058 2.90967770
#>  [2,] 0.600563705 9.12840221 0.29932925
#>  [3,] 0.210440720 0.03262397 0.09288229
#>  [4,] 0.989052146 3.01334736 0.24112646
#>  [5,] 0.144300272 0.32482088 1.22241130
#>  [6,] 0.254315338 0.59383347 0.30298859
#>  [7,] 0.425865217 1.58950797 0.88262488
#>  [8,] 4.415563634 0.02177716 0.42309449
#>  [9,] 0.021541979 1.47831253 2.09762025
#> [10,] 2.257321777 1.70762265 1.21493698
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5551057 0.5551057 0.5551057
#>  [2,] 0.9808685 0.9808685 0.9808685
#>  [3,] 1.2939071 1.2939071 1.2939071
#>  [4,] 0.2093374 0.2093374 0.2093374
#>  [5,] 1.6594480 1.6594480 1.6594480
#>  [6,] 2.1319380 2.1319380 2.1319380
#>  [7,] 0.6311066 0.6311066 0.6311066
#>  [8,] 1.5854341 1.5854341 1.5854341
#>  [9,] 0.2704854 0.2704854 0.2704854
#> [10,] 1.3309397 1.3309397 1.3309397

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,]  1.2125978 2.6778283 0.72269925
#>  [2,] 10.5076673 0.9705976 0.26551233
#>  [3,]  1.6500358 1.6846869 4.36011118
#>  [4,]  0.9005350 0.9005350 0.90053500
#>  [5,]  3.8467060 0.3799720 0.37997200
#>  [6,]  0.5575484 4.2995424 0.08700868
#>  [7,]  2.6131980 1.2855666 2.32689000
#>  [8,]  1.5847630 1.5390902 1.53909021
#>  [9,]  1.8501592 1.8501592 4.16723287
#> [10,]  2.9756224 4.4027325 7.90263995
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08880639 3.78249758 0.26059842
#>  [2,] 0.32685792 0.59295043 0.03034445
#>  [3,] 0.94642561 1.21815495 1.53887309
#>  [4,] 0.12188658 0.64774175 0.33898454
#>  [5,] 0.40499034 0.34644626 0.52604576
#>  [6,] 0.57876412 0.08911526 0.55590106
#>  [7,] 0.25622191 3.21836331 0.02811806
#>  [8,] 3.84299754 0.27608961 2.48189769
#>  [9,] 3.00511532 0.84574603 3.52509553
#> [10,] 0.45434859 0.89469957 1.27066758
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.43325076 0.43325076 0.43325076
#>  [2,] 0.02589083 0.02589083 0.02589083
#>  [3,] 0.22641801 0.22641801 0.22641801
#>  [4,] 0.47101275 0.47101275 0.47101275
#>  [5,] 3.02078985 3.02078985 3.02078985
#>  [6,] 0.81592507 0.81592507 0.81592507
#>  [7,] 0.42676398 0.42676398 0.42676398
#>  [8,] 0.56479166 0.56479166 0.56479166
#>  [9,] 1.09265451 1.09265451 1.09265451
#> [10,] 0.76838172 0.76838172 0.76838172