Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.7570584 0.4607102 0.4607102
#>  [2,] 0.6583441 2.2633275 1.9254033
#>  [3,] 0.8000522 0.4477182 0.8000522
#>  [4,] 1.4347096 1.5369480 1.8541204
#>  [5,] 1.6225604 1.6225604 1.0611143
#>  [6,] 0.4468450 0.4468450 1.4497485
#>  [7,] 0.3647019 0.7421941 0.3400279
#>  [8,] 0.2446535 0.3885835 0.3885835
#>  [9,] 0.2869271 0.2869271 0.2869271
#> [10,] 0.7593272 0.6646905 0.6646905
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.73314992 0.5934742 0.28388329
#>  [2,] 1.17140475 0.3852223 0.32852126
#>  [3,] 1.88523636 1.2368773 0.08638464
#>  [4,] 0.80897532 0.4659788 0.26311047
#>  [5,] 0.04583994 1.4606745 2.11176227
#>  [6,] 2.73570170 0.9741193 1.14083772
#>  [7,] 1.74483771 0.1961893 1.67802917
#>  [8,] 0.25246829 0.2294651 0.14455781
#>  [9,] 0.04876110 0.9461374 3.67979823
#> [10,] 0.76125179 2.7608619 1.57069490
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02372456 0.02372456 0.02372456
#>  [2,] 1.70837114 1.70837114 1.70837114
#>  [3,] 0.19996609 0.19996609 0.19996609
#>  [4,] 1.89300097 1.89300097 1.89300097
#>  [5,] 0.77348107 0.77348107 0.77348107
#>  [6,] 0.86267590 0.86267590 0.86267590
#>  [7,] 0.40513295 0.40513295 0.40513295
#>  [8,] 0.49816732 0.49816732 0.49816732
#>  [9,] 0.05583675 0.05583675 0.05583675
#> [10,] 0.85065157 0.85065157 0.85065157

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.93979041 1.47654430 1.83249203
#>  [2,] 0.29156356 2.89073713 4.88190345
#>  [3,] 1.88179230 0.55120945 1.16482147
#>  [4,] 0.07696705 0.07696705 0.07696705
#>  [5,] 0.78326419 2.66364250 2.66364250
#>  [6,] 0.25322625 0.25322625 0.25322625
#>  [7,] 2.40503297 0.58945879 2.40503297
#>  [8,] 0.34476287 2.87839795 2.87839795
#>  [9,] 0.59594282 0.59594282 0.59594282
#> [10,] 1.21571201 1.48374700 1.48374700
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 0.3946361 0.944203987 0.06926981
#>  [2,] 0.4921712 1.495523077 1.51207222
#>  [3,] 1.0942658 2.255257530 0.43626562
#>  [4,] 0.7043374 1.736352182 0.38126085
#>  [5,] 0.6051175 0.860486435 3.24298392
#>  [6,] 0.7881268 0.336547168 0.93447799
#>  [7,] 1.1435302 1.575419074 0.59539730
#>  [8,] 0.3998422 1.002556984 0.73563862
#>  [9,] 3.3470445 0.007101348 0.46140575
#> [10,] 0.8614209 2.406859515 1.01320402
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5526360 0.5526360 0.5526360
#>  [2,] 0.1873734 0.1873734 0.1873734
#>  [3,] 2.0342621 2.0342621 2.0342621
#>  [4,] 0.9785232 0.9785232 0.9785232
#>  [5,] 0.6819525 0.6819525 0.6819525
#>  [6,] 0.7036793 0.7036793 0.7036793
#>  [7,] 1.6864082 1.6864082 1.6864082
#>  [8,] 1.8690029 1.8690029 1.8690029
#>  [9,] 0.9482212 0.9482212 0.9482212
#> [10,] 1.1516062 1.1516062 1.1516062

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.29757599 0.6592146 2.2870028
#>  [2,] 0.65295291 0.6529529 0.3679259
#>  [3,] 0.01044761 2.1189271 0.5176081
#>  [4,] 0.09063207 0.4243896 0.4243896
#>  [5,] 0.43300846 0.5332480 2.8526034
#>  [6,] 5.38916311 2.6144080 6.0501042
#>  [7,] 0.10112937 0.1011294 0.1011294
#>  [8,] 2.37484354 4.5484454 2.3748435
#>  [9,] 0.27055437 0.7211833 0.7211833
#> [10,] 1.27599787 1.2759979 1.2759979
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>               [,1]       [,2]      [,3]
#>  [1,] 0.2713897751 0.21698901 2.3986578
#>  [2,] 2.0853357686 0.30474165 0.9445736
#>  [3,] 0.9475544020 0.65362858 1.1220975
#>  [4,] 0.9367693630 0.87415785 1.9159261
#>  [5,] 0.0139758133 0.34088548 0.3150828
#>  [6,] 0.6702401978 0.08914259 1.8055032
#>  [7,] 0.1826944822 2.53524629 0.4435867
#>  [8,] 0.0009955576 0.11117384 2.7512407
#>  [9,] 0.5105280586 1.31828682 0.5802264
#> [10,] 0.1303157826 0.08076415 0.2209822
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4740063 0.4740063 0.4740063
#>  [2,] 0.3031253 0.3031253 0.3031253
#>  [3,] 0.6693183 0.6693183 0.6693183
#>  [4,] 3.3998689 3.3998689 3.3998689
#>  [5,] 1.0263654 1.0263654 1.0263654
#>  [6,] 1.4420080 1.4420080 1.4420080
#>  [7,] 0.2215211 0.2215211 0.2215211
#>  [8,] 0.5817400 0.5817400 0.5817400
#>  [9,] 0.4392750 0.4392750 0.4392750
#> [10,] 0.1750266 0.1750266 0.1750266