Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2379723 0.52436480 0.2379723
#>  [2,] 1.3274876 0.31810833 1.9757044
#>  [3,] 1.4523552 1.45235522 1.3691082
#>  [4,] 1.0791188 1.07911877 1.0791188
#>  [5,] 0.5678236 0.02823694 0.1668903
#>  [6,] 1.0425764 1.16052242 1.1605224
#>  [7,] 1.4234910 0.03532627 1.4234910
#>  [8,] 0.2589066 2.31613567 0.2589066
#>  [9,] 0.7109581 1.55320352 1.3437499
#> [10,] 1.1458735 0.48581397 3.5282482
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3618848 0.48455573 2.88258893
#>  [2,] 0.7925943 0.29777850 0.05525133
#>  [3,] 4.4470264 0.02712424 0.37672550
#>  [4,] 0.3666379 0.03974344 0.53443271
#>  [5,] 0.9287976 0.15141093 4.12715177
#>  [6,] 0.5350817 0.39689401 3.10080622
#>  [7,] 0.2234086 1.04839819 1.79163739
#>  [8,] 0.1394703 4.77671702 1.04303379
#>  [9,] 0.7996736 0.56313821 1.01349467
#> [10,] 0.2048247 1.51394024 1.08024291
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 4.15120674 4.15120674 4.15120674
#>  [2,] 0.26918452 0.26918452 0.26918452
#>  [3,] 2.29298284 2.29298284 2.29298284
#>  [4,] 1.72457090 1.72457090 1.72457090
#>  [5,] 0.81844000 0.81844000 0.81844000
#>  [6,] 0.32637792 0.32637792 0.32637792
#>  [7,] 1.84219666 1.84219666 1.84219666
#>  [8,] 0.19183480 0.19183480 0.19183480
#>  [9,] 0.08917456 0.08917456 0.08917456
#> [10,] 1.67316359 1.67316359 1.67316359

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>               [,1]      [,2]         [,3]
#>  [1,] 0.2496589521 0.2496590 0.2496589521
#>  [2,] 1.0852484899 1.1050750 0.0289213151
#>  [3,] 0.0003710063 0.3305257 0.0003710063
#>  [4,] 1.3242558250 1.8481968 1.0079931491
#>  [5,] 1.9354843859 0.3110649 0.6758340879
#>  [6,] 0.9523431361 0.3495747 0.3495746987
#>  [7,] 0.1283837105 0.3789439 0.3789439215
#>  [8,] 0.8998382128 0.8998382 2.2709440672
#>  [9,] 1.7165031005 0.8811724 2.1582331644
#> [10,] 0.5948330731 0.5109766 0.5948330731
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2759038 2.0182388 0.53037109
#>  [2,] 1.6946851 0.7004330 3.75451375
#>  [3,] 0.2404467 0.2467060 0.89562716
#>  [4,] 0.7166292 0.5827093 1.31120375
#>  [5,] 1.3376578 0.3579659 0.13476640
#>  [6,] 1.0066129 0.2232767 0.07923683
#>  [7,] 1.7179745 0.2556354 0.95453105
#>  [8,] 2.1024311 1.7061711 0.70074733
#>  [9,] 0.6331914 0.9981201 0.37731278
#> [10,] 0.3196067 0.8010288 0.08774382
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.94268130 0.94268130 0.94268130
#>  [2,] 0.82169449 0.82169449 0.82169449
#>  [3,] 0.19125994 0.19125994 0.19125994
#>  [4,] 2.44517095 2.44517095 2.44517095
#>  [5,] 2.31560935 2.31560935 2.31560935
#>  [6,] 0.06604705 0.06604705 0.06604705
#>  [7,] 0.37410617 0.37410617 0.37410617
#>  [8,] 1.33888502 1.33888502 1.33888502
#>  [9,] 1.53658251 1.53658251 1.53658251
#> [10,] 0.86320527 0.86320527 0.86320527

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 3.97570438 0.99641524 2.63478876
#>  [2,] 0.03692759 0.03692759 0.03692759
#>  [3,] 0.90744346 0.90744346 0.90646165
#>  [4,] 0.13010966 0.13010966 0.13010966
#>  [5,] 1.14167757 0.17296960 0.21667480
#>  [6,] 2.00234088 1.14024221 1.64375027
#>  [7,] 1.28521279 1.22782636 1.28521279
#>  [8,] 0.96687104 4.77588650 2.58732279
#>  [9,] 0.90882474 2.92996723 1.23205167
#> [10,] 1.63891188 1.08806753 1.08806753
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2147582 2.50586061 1.3203948
#>  [2,] 0.3600440 0.10771472 1.8113207
#>  [3,] 0.9616431 0.90857638 2.3655227
#>  [4,] 0.4592367 0.94639793 4.1555587
#>  [5,] 2.3203379 0.34641404 0.5206122
#>  [6,] 1.5791158 0.06479894 0.1331658
#>  [7,] 0.4481423 1.30756271 4.5333637
#>  [8,] 0.4837481 1.69470397 0.2624538
#>  [9,] 2.6384410 0.69037097 0.3390825
#> [10,] 1.1939404 0.01235186 2.0937150
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42722965 0.42722965 0.42722965
#>  [2,] 0.32668995 0.32668995 0.32668995
#>  [3,] 0.03052432 0.03052432 0.03052432
#>  [4,] 2.34144079 2.34144079 2.34144079
#>  [5,] 1.52666453 1.52666453 1.52666453
#>  [6,] 0.07054027 0.07054027 0.07054027
#>  [7,] 0.62671096 0.62671096 0.62671096
#>  [8,] 1.74546765 1.74546765 1.74546765
#>  [9,] 0.57696308 0.57696308 0.57696308
#> [10,] 0.32535232 0.32535232 0.32535232