Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27960322 0.76503966 0.76503966
#>  [2,] 0.48609813 0.48609813 0.48609813
#>  [3,] 0.09304513 0.71194522 0.09304513
#>  [4,] 0.84102858 0.28972862 0.28972862
#>  [5,] 0.63314570 0.63314570 0.63314570
#>  [6,] 0.01499713 1.00711556 0.97473488
#>  [7,] 0.17673616 0.17673616 0.17673616
#>  [8,] 1.91676306 1.82085451 1.07709681
#>  [9,] 0.58931946 0.02443347 0.02443347
#> [10,] 3.34905469 0.50539491 0.46300259
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.7307215 1.09113210 2.763383405
#>  [2,] 4.8568441 0.25720615 0.194060490
#>  [3,] 2.5435365 1.44830870 0.113834966
#>  [4,] 0.1405732 1.13725182 0.331424856
#>  [5,] 0.2953873 0.14372980 0.005656142
#>  [6,] 1.0091233 0.02542221 0.239146998
#>  [7,] 0.2496123 2.69511221 0.479484779
#>  [8,] 0.1527490 0.15044224 0.362795558
#>  [9,] 0.2746649 0.77540225 5.272563576
#> [10,] 0.9000029 1.40077348 1.305749935
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.005416118 0.005416118 0.005416118
#>  [2,] 1.021879144 1.021879144 1.021879144
#>  [3,] 0.241637159 0.241637159 0.241637159
#>  [4,] 0.928730628 0.928730628 0.928730628
#>  [5,] 0.163942592 0.163942592 0.163942592
#>  [6,] 0.133599546 0.133599546 0.133599546
#>  [7,] 0.370550567 0.370550567 0.370550567
#>  [8,] 0.438450224 0.438450224 0.438450224
#>  [9,] 0.791613751 0.791613751 0.791613751
#> [10,] 1.724351071 1.724351071 1.724351071

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.847375570 0.84737557 1.05245107
#>  [2,] 0.121578004 0.12157800 0.12157800
#>  [3,] 0.056933568 1.05676924 0.50727433
#>  [4,] 0.316133070 0.87840511 1.08644938
#>  [5,] 0.008795495 2.98428401 1.43281294
#>  [6,] 0.058104367 0.05810437 0.05810437
#>  [7,] 0.540077368 0.54007737 0.54007737
#>  [8,] 0.507247425 0.57057910 0.50724742
#>  [9,] 1.243054521 0.76399778 0.76399778
#> [10,] 1.015681882 0.26689113 1.01568188
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.82543886 0.2998561 1.20912203
#>  [2,] 0.84919134 0.0207160 0.93240262
#>  [3,] 0.32789288 2.0539585 0.71710148
#>  [4,] 0.04093305 0.5388245 0.69925016
#>  [5,] 1.22859177 2.2578288 0.56483941
#>  [6,] 0.53600984 0.0721111 0.02948869
#>  [7,] 0.29321056 1.1061916 1.16885880
#>  [8,] 0.88903520 1.0296251 1.14806938
#>  [9,] 2.01257305 0.6071345 2.86522059
#> [10,] 0.44077816 0.9033709 0.39021176
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.37232753 2.37232753 2.37232753
#>  [2,] 0.14108509 0.14108509 0.14108509
#>  [3,] 1.16386643 1.16386643 1.16386643
#>  [4,] 0.07317221 0.07317221 0.07317221
#>  [5,] 0.19813172 0.19813172 0.19813172
#>  [6,] 1.53296637 1.53296637 1.53296637
#>  [7,] 1.79644967 1.79644967 1.79644967
#>  [8,] 0.37573262 0.37573262 0.37573262
#>  [9,] 7.45963453 7.45963453 7.45963453
#> [10,] 0.38123766 0.38123766 0.38123766

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 3.6617154 6.8813221 7.3456553
#>  [2,] 1.2223496 1.2223496 0.1355312
#>  [3,] 1.0550158 1.0550158 1.0441900
#>  [4,] 0.2687357 0.4801970 0.4801970
#>  [5,] 2.7382324 2.2245683 0.9524648
#>  [6,] 0.7472763 1.1492893 0.3645081
#>  [7,] 0.3554024 0.3554024 0.3554024
#>  [8,] 2.7145046 3.3718110 3.3718110
#>  [9,] 0.5202208 0.3919429 0.5202208
#> [10,] 1.0543603 1.0543603 1.0543603
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9957608 1.5103646 3.25345459
#>  [2,] 0.2154627 0.5613635 2.04970466
#>  [3,] 0.3708979 1.2551013 0.56952363
#>  [4,] 0.1804144 0.6999505 0.05906389
#>  [5,] 0.4901722 2.7975759 1.87132927
#>  [6,] 1.6100265 0.1276632 2.26114610
#>  [7,] 3.8927660 0.1860355 0.74665625
#>  [8,] 0.4534944 0.4698799 1.04521595
#>  [9,] 2.1176740 2.1372184 0.57018081
#> [10,] 0.6182660 1.2953874 0.51828102
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2613244 0.2613244 0.2613244
#>  [2,] 0.8451327 0.8451327 0.8451327
#>  [3,] 0.5471952 0.5471952 0.5471952
#>  [4,] 1.1802422 1.1802422 1.1802422
#>  [5,] 1.3375466 1.3375466 1.3375466
#>  [6,] 1.2205718 1.2205718 1.2205718
#>  [7,] 3.7620613 3.7620613 3.7620613
#>  [8,] 0.1330684 0.1330684 0.1330684
#>  [9,] 1.2424846 1.2424846 1.2424846
#> [10,] 1.1141587 1.1141587 1.1141587