Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60347113 0.46362554 0.44014326
#>  [2,] 1.76065752 1.76065752 1.40863363
#>  [3,] 0.02742629 0.69253917 0.69253917
#>  [4,] 1.75532428 1.48798968 3.79938432
#>  [5,] 3.57167412 0.25800718 0.25800718
#>  [6,] 0.87659330 0.59995233 0.59995233
#>  [7,] 2.65851881 4.98524255 0.78472941
#>  [8,] 1.49588863 0.32831733 0.93445867
#>  [9,] 0.20667120 0.20667120 0.20667120
#> [10,] 0.01402557 0.01402557 0.01402557
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2129236 0.2666525 0.40609374
#>  [2,] 0.3462879 0.0383289 0.84968001
#>  [3,] 0.5274446 0.9672570 0.51562791
#>  [4,] 2.1418789 2.6008307 0.28496380
#>  [5,] 0.9006775 0.3256038 0.44078159
#>  [6,] 0.9370740 1.0325399 0.69468051
#>  [7,] 0.4639561 0.0707166 0.28746496
#>  [8,] 3.6168183 0.5057292 0.62914459
#>  [9,] 2.9599698 3.7520089 0.08719172
#> [10,] 1.5396865 0.4570073 2.10834571
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.83812090 0.83812090 0.83812090
#>  [2,] 3.44812176 3.44812176 3.44812176
#>  [3,] 0.11210643 0.11210643 0.11210643
#>  [4,] 0.11345883 0.11345883 0.11345883
#>  [5,] 0.08768564 0.08768564 0.08768564
#>  [6,] 0.03378196 0.03378196 0.03378196
#>  [7,] 0.09797945 0.09797945 0.09797945
#>  [8,] 0.45962178 0.45962178 0.45962178
#>  [9,] 0.73513691 0.73513691 0.73513691
#> [10,] 0.69680583 0.69680583 0.69680583

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4473954 0.4473954 0.4473954
#>  [2,] 1.1520750 2.7506143 3.0578732
#>  [3,] 0.9522195 4.5542401 0.3280069
#>  [4,] 0.2448738 2.2879077 1.7669245
#>  [5,] 0.6949606 0.1510265 0.4541287
#>  [6,] 0.5937222 0.5937222 1.0751584
#>  [7,] 0.6315045 0.6079653 0.6315045
#>  [8,] 0.9793446 1.3512910 1.3512910
#>  [9,] 0.4606219 2.2056838 1.5904276
#> [10,] 1.6036778 2.3023110 2.5104752
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.50464788 0.3217000 3.2003875
#>  [2,] 0.48547393 3.5507347 0.8216558
#>  [3,] 0.46514028 1.2228241 0.4778631
#>  [4,] 1.94384605 1.7054809 0.6869118
#>  [5,] 0.11264584 2.1354707 0.1566493
#>  [6,] 0.07771925 0.4777905 0.1006721
#>  [7,] 0.55292994 0.7047266 2.5919676
#>  [8,] 0.65867054 0.1724296 3.7342118
#>  [9,] 1.58408102 0.7284670 0.3600407
#> [10,] 1.46421942 1.0135511 1.9496128
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.77186874 0.77186874 0.77186874
#>  [2,] 0.37798760 0.37798760 0.37798760
#>  [3,] 2.03371054 2.03371054 2.03371054
#>  [4,] 0.22673872 0.22673872 0.22673872
#>  [5,] 0.87933889 0.87933889 0.87933889
#>  [6,] 1.11062119 1.11062119 1.11062119
#>  [7,] 0.56337135 0.56337135 0.56337135
#>  [8,] 0.03306882 0.03306882 0.03306882
#>  [9,] 0.64519736 0.64519736 0.64519736
#> [10,] 0.20185204 0.20185204 0.20185204

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 7.45111008 0.17100812 0.34945089
#>  [2,] 1.14381576 1.60370521 0.70612622
#>  [3,] 5.59467804 1.61530330 0.06212277
#>  [4,] 2.91532416 0.04866273 0.62214802
#>  [5,] 0.74311801 0.35223723 0.74311801
#>  [6,] 0.38172235 0.44773187 0.44773187
#>  [7,] 1.42709242 2.58819548 0.19956097
#>  [8,] 0.44100493 0.44100493 0.44100493
#>  [9,] 0.03611317 0.03611317 0.03611317
#> [10,] 0.50634902 0.50634902 0.50634902
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]        [,2]       [,3]
#>  [1,] 1.2423955 0.009375938 2.85139223
#>  [2,] 1.5788291 0.324977453 0.45484327
#>  [3,] 1.4865409 3.770640822 0.70833584
#>  [4,] 0.6217524 0.010818256 0.14810986
#>  [5,] 3.0513937 4.913420388 2.48855034
#>  [6,] 5.6895697 0.134580428 1.00632759
#>  [7,] 0.3150028 0.500708187 0.06799831
#>  [8,] 1.2239459 0.426973651 2.22179237
#>  [9,] 0.3217666 1.797251080 1.16066385
#> [10,] 0.3723108 1.339628125 2.90845603
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.24418232 1.24418232 1.24418232
#>  [2,] 0.03410061 0.03410061 0.03410061
#>  [3,] 1.24766661 1.24766661 1.24766661
#>  [4,] 3.10136813 3.10136813 3.10136813
#>  [5,] 0.05181178 0.05181178 0.05181178
#>  [6,] 1.17077546 1.17077546 1.17077546
#>  [7,] 0.02303899 0.02303899 0.02303899
#>  [8,] 3.52991723 3.52991723 3.52991723
#>  [9,] 2.30109527 2.30109527 2.30109527
#> [10,] 1.95009474 1.95009474 1.95009474