Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5596232 0.3966654 1.6994009
#>  [2,] 0.0782184 0.0782184 0.0782184
#>  [3,] 1.0331031 1.0331031 1.0331031
#>  [4,] 0.2730956 0.2730956 0.2328417
#>  [5,] 0.9530955 0.7507278 1.0184550
#>  [6,] 0.8249219 0.4721933 0.7780737
#>  [7,] 0.4301278 1.1623925 1.1623925
#>  [8,] 0.4923259 0.1791571 0.1791571
#>  [9,] 0.5906931 0.6471002 0.6471002
#> [10,] 2.4541522 2.1629276 1.2719878
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.43673844 0.006533659 0.73127506
#>  [2,] 0.42970594 2.433528150 0.30077917
#>  [3,] 0.33698353 0.501079078 0.75234104
#>  [4,] 0.19063405 0.341442076 0.75327600
#>  [5,] 0.18312211 2.467066864 0.61083183
#>  [6,] 0.06252457 0.923557197 1.50873485
#>  [7,] 0.72368884 2.379271664 0.31325453
#>  [8,] 0.97463598 2.380319363 0.74652338
#>  [9,] 3.88330562 0.678249591 4.38411265
#> [10,] 0.24473966 0.331860193 0.06105926
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.709063383 1.709063383 1.709063383
#>  [2,] 2.065898625 2.065898625 2.065898625
#>  [3,] 0.390233914 0.390233914 0.390233914
#>  [4,] 1.238932516 1.238932516 1.238932516
#>  [5,] 1.242995100 1.242995100 1.242995100
#>  [6,] 0.398084263 0.398084263 0.398084263
#>  [7,] 0.001369525 0.001369525 0.001369525
#>  [8,] 4.302574878 4.302574878 4.302574878
#>  [9,] 0.788579928 0.788579928 0.788579928
#> [10,] 2.097985079 2.097985079 2.097985079

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.74515003 3.9553615 1.4236572
#>  [2,] 1.14124538 2.6580421 1.1412454
#>  [3,] 0.04255348 0.1185413 1.7173825
#>  [4,] 0.02325169 0.6364281 0.8929998
#>  [5,] 1.43788148 1.4378815 1.4378815
#>  [6,] 2.03143249 4.2675942 4.9236247
#>  [7,] 1.57743664 1.5774366 1.5774366
#>  [8,] 0.65666309 2.2326082 2.2326082
#>  [9,] 0.39993573 0.3999357 0.3999357
#> [10,] 0.20115115 1.2969668 1.2969668
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]        [,2]      [,3]
#>  [1,] 0.26471078 1.437707605 1.9087155
#>  [2,] 1.22884547 1.370288813 1.0246210
#>  [3,] 0.04333208 0.869941177 0.3737676
#>  [4,] 2.10240265 0.608498990 0.1602788
#>  [5,] 0.31766585 1.144751475 0.6328010
#>  [6,] 2.05488090 0.925256895 4.6311599
#>  [7,] 0.68605245 0.848053078 0.5398955
#>  [8,] 0.59446550 0.628861320 0.3841638
#>  [9,] 0.55880044 0.006490818 1.9772119
#> [10,] 1.07558628 0.813635141 2.3436266
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.896593578 1.896593578 1.896593578
#>  [2,] 0.101278315 0.101278315 0.101278315
#>  [3,] 0.003998997 0.003998997 0.003998997
#>  [4,] 0.372556383 0.372556383 0.372556383
#>  [5,] 0.788618394 0.788618394 0.788618394
#>  [6,] 1.000330309 1.000330309 1.000330309
#>  [7,] 0.090362091 0.090362091 0.090362091
#>  [8,] 0.133548179 0.133548179 0.133548179
#>  [9,] 1.236941571 1.236941571 1.236941571
#> [10,] 0.356469204 0.356469204 0.356469204

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.24023894 2.8356177 1.87965754
#>  [2,] 0.26668097 0.7561743 0.75617431
#>  [3,] 0.46057348 0.3933302 0.51578601
#>  [4,] 0.76820574 0.7868769 0.48760304
#>  [5,] 0.85584543 1.9888495 1.98884953
#>  [6,] 0.06984958 0.7082598 0.06984958
#>  [7,] 0.40213060 0.4021306 0.39671529
#>  [8,] 1.14910659 0.2697344 5.18619121
#>  [9,] 1.15867031 0.5408393 0.54083931
#> [10,] 1.21698368 1.2169837 4.14603565
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 2.74275177 0.07194351 2.2024848
#>  [2,] 0.67616849 2.09268741 1.2795502
#>  [3,] 0.10657338 0.28831627 0.5937283
#>  [4,] 1.05483619 1.43048062 0.6032161
#>  [5,] 0.06955227 1.29194866 0.2438875
#>  [6,] 2.98337743 0.79544796 0.8157414
#>  [7,] 0.20156948 0.97886299 1.6189804
#>  [8,] 0.02336862 0.76314577 0.9868137
#>  [9,] 0.53505007 1.23305363 0.7466867
#> [10,] 0.95846722 5.27034506 1.1501850
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.64199341 0.64199341 0.64199341
#>  [2,] 0.17236773 0.17236773 0.17236773
#>  [3,] 1.92381051 1.92381051 1.92381051
#>  [4,] 1.21211445 1.21211445 1.21211445
#>  [5,] 0.05447654 0.05447654 0.05447654
#>  [6,] 0.85321527 0.85321527 0.85321527
#>  [7,] 0.73868380 0.73868380 0.73868380
#>  [8,] 1.15683473 1.15683473 1.15683473
#>  [9,] 0.06197263 0.06197263 0.06197263
#> [10,] 0.28301092 0.28301092 0.28301092