Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11883639 0.11883639 1.02641319
#>  [2,] 1.31521046 1.38025238 0.33473010
#>  [3,] 0.89069738 0.89069738 0.02938749
#>  [4,] 2.61610175 2.61610175 2.61610175
#>  [5,] 0.07445913 0.09464202 0.09104962
#>  [6,] 1.19712960 0.31069638 1.56133073
#>  [7,] 0.70719897 0.70719897 0.70719897
#>  [8,] 0.30586997 0.30586997 0.82464784
#>  [9,] 0.46810558 0.92618612 0.46439365
#> [10,] 0.33305158 0.33305158 0.33305158
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6439382 0.97404244 1.7610537
#>  [2,] 0.1274783 1.18933401 0.9998853
#>  [3,] 0.1902357 0.03902921 1.4328113
#>  [4,] 0.7853126 1.13250259 0.1132832
#>  [5,] 0.7114690 1.64185779 0.5796107
#>  [6,] 1.1905179 0.39200875 0.9809989
#>  [7,] 0.6150720 0.27174504 0.6479812
#>  [8,] 0.6680010 0.00363665 2.5415270
#>  [9,] 0.2319047 1.56520551 1.1664935
#> [10,] 0.6028718 0.17260469 0.3685466
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.496206771 0.496206771 0.496206771
#>  [2,] 1.139079854 1.139079854 1.139079854
#>  [3,] 0.460643921 0.460643921 0.460643921
#>  [4,] 1.864541046 1.864541046 1.864541046
#>  [5,] 1.086133520 1.086133520 1.086133520
#>  [6,] 0.023729460 0.023729460 0.023729460
#>  [7,] 0.007869401 0.007869401 0.007869401
#>  [8,] 1.019611146 1.019611146 1.019611146
#>  [9,] 1.796378517 1.796378517 1.796378517
#> [10,] 0.246677901 0.246677901 0.246677901

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4332470 2.4687964 3.99433750
#>  [2,] 2.8098003 2.8098003 2.80980026
#>  [3,] 0.1313166 3.1002121 0.09370294
#>  [4,] 0.4559586 0.1476729 0.45595860
#>  [5,] 1.4281507 1.4281507 0.17664760
#>  [6,] 0.6163083 0.7868769 0.61630828
#>  [7,] 0.3353955 0.4276784 0.42767843
#>  [8,] 1.2185622 1.2185622 1.21856220
#>  [9,] 3.6260300 4.0059412 3.21830477
#> [10,] 1.6393299 0.4763021 1.63932987
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]       [,2]        [,3]
#>  [1,] 0.9480006 5.19934927 0.257509968
#>  [2,] 0.2817066 0.67300004 1.034295845
#>  [3,] 1.0274065 0.04530776 0.482669150
#>  [4,] 3.1109813 0.36731378 0.649024790
#>  [5,] 5.5557736 1.24367648 0.008877895
#>  [6,] 1.0123050 0.74346013 0.141472437
#>  [7,] 0.4360681 1.70762265 0.890722599
#>  [8,] 0.0503070 0.71750950 0.837053468
#>  [9,] 1.4842117 2.22834874 0.789976380
#> [10,] 0.6908985 0.47007352 3.419427663
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02154817 0.02154817 0.02154817
#>  [2,] 0.95802935 0.95802935 0.95802935
#>  [3,] 0.15460969 0.15460969 0.15460969
#>  [4,] 0.61884896 0.61884896 0.61884896
#>  [5,] 3.28294802 3.28294802 3.28294802
#>  [6,] 0.11947773 0.11947773 0.11947773
#>  [7,] 0.33797228 0.33797228 0.33797228
#>  [8,] 0.26668841 0.26668841 0.26668841
#>  [9,] 0.73042807 0.73042807 0.73042807
#> [10,] 1.11427172 1.11427172 1.11427172

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1784796 0.2689083 3.21535925
#>  [2,] 0.2174689 0.1360774 0.21746894
#>  [3,] 0.9534623 0.9534623 0.95346227
#>  [4,] 0.8560074 0.3184814 0.85600743
#>  [5,] 1.3135027 1.3135027 4.65699087
#>  [6,] 0.5852844 0.9421370 0.93123694
#>  [7,] 0.7040563 0.7040563 1.29575685
#>  [8,] 0.4233155 0.4233155 0.42331549
#>  [9,] 0.8137158 1.3436686 0.08474568
#> [10,] 0.5783777 1.4943051 1.49430507
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.7672893 0.66802259 1.30485596
#>  [2,] 0.9816598 0.24552734 2.22712487
#>  [3,] 1.1393010 0.21056042 1.14676694
#>  [4,] 2.9222742 0.30496272 0.50489582
#>  [5,] 0.9401959 0.28048260 0.10151122
#>  [6,] 1.0341698 0.31456317 5.07376240
#>  [7,] 0.1116165 0.10601681 0.61377641
#>  [8,] 3.5178029 0.08600592 0.57036318
#>  [9,] 0.4351192 0.01917451 0.02766223
#> [10,] 2.4538676 1.51862236 1.62935288
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.85637130 1.85637130 1.85637130
#>  [2,] 1.37738638 1.37738638 1.37738638
#>  [3,] 0.23513750 0.23513750 0.23513750
#>  [4,] 0.09537849 0.09537849 0.09537849
#>  [5,] 0.42936729 0.42936729 0.42936729
#>  [6,] 1.25645049 1.25645049 1.25645049
#>  [7,] 1.34529330 1.34529330 1.34529330
#>  [8,] 1.62739354 1.62739354 1.62739354
#>  [9,] 0.49911981 0.49911981 0.49911981
#> [10,] 1.27907302 1.27907302 1.27907302