Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensitiesintensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>            [,1]      [,2]        [,3]
#>  [1,] 2.6163881 2.1311581 1.278147732
#>  [2,] 0.4085450 0.4085450 0.003170595
#>  [3,] 1.6999473 1.6999473 1.540347270
#>  [4,] 2.3318052 3.6468928 2.331805227
#>  [5,] 0.3515241 2.0923913 2.063504284
#>  [6,] 0.7293732 0.7293732 2.655284980
#>  [7,] 0.2515103 1.0151643 0.743277220
#>  [8,] 0.6928676 0.6243192 0.692867617
#>  [9,] 0.9346046 2.7106074 0.575292389
#> [10,] 0.6281616 0.4936803 0.628161570
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>              [,1]      [,2]       [,3]
#>  [1,] 1.546512803 1.6203734 0.33061792
#>  [2,] 1.589744943 1.1798863 2.36973398
#>  [3,] 0.004202258 0.9971734 1.40346816
#>  [4,] 0.068782028 0.0105614 0.21346916
#>  [5,] 0.156154942 2.8550510 0.37824424
#>  [6,] 1.140089448 1.0799607 0.02168446
#>  [7,] 0.009951358 0.6373465 0.71566503
#>  [8,] 0.335220305 3.0671916 0.28549873
#>  [9,] 0.057183482 0.2842011 0.89729089
#> [10,] 0.458357729 0.6933708 0.40156580
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32482992 0.32482992 0.32482992
#>  [2,] 0.98032146 0.98032146 0.98032146
#>  [3,] 0.15144278 0.15144278 0.15144278
#>  [4,] 0.72111765 0.72111765 0.72111765
#>  [5,] 0.30259918 0.30259918 0.30259918
#>  [6,] 0.82203079 0.82203079 0.82203079
#>  [7,] 1.11147800 1.11147800 1.11147800
#>  [8,] 0.02366681 0.02366681 0.02366681
#>  [9,] 3.02277954 3.02277954 3.02277954
#> [10,] 2.21821757 2.21821757 2.21821757

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.15522315 1.15522315 0.46863343
#>  [2,] 0.48845315 0.41304923 0.41304923
#>  [3,] 0.24100912 1.14847758 0.37056447
#>  [4,] 0.06619637 0.06619637 0.06619637
#>  [5,] 2.61257519 0.34451051 3.82145525
#>  [6,] 0.38970156 0.48994028 0.38970156
#>  [7,] 0.35247409 1.09000928 0.35247409
#>  [8,] 0.67017087 0.63199835 0.63199835
#>  [9,] 0.02287897 0.02287897 0.02287897
#> [10,] 0.16727037 0.24213131 0.16727037
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.68329455 0.58509556 0.07977899
#>  [2,] 1.63141564 1.16171193 1.75283562
#>  [3,] 0.08004267 0.06292968 1.49805714
#>  [4,] 0.33862339 1.08792749 0.62217442
#>  [5,] 0.13472576 0.40318663 1.10511411
#>  [6,] 0.08376555 2.29632153 1.13824398
#>  [7,] 0.30958062 1.37062467 0.28501239
#>  [8,] 1.42849108 1.24640147 0.20116612
#>  [9,] 0.11145813 0.28756832 0.83001565
#> [10,] 1.30717563 0.34123809 0.43025518
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7140510 0.7140510 0.7140510
#>  [2,] 0.9745535 0.9745535 0.9745535
#>  [3,] 0.9429402 0.9429402 0.9429402
#>  [4,] 3.8619131 3.8619131 3.8619131
#>  [5,] 0.6633373 0.6633373 0.6633373
#>  [6,] 1.3454606 1.3454606 1.3454606
#>  [7,] 0.5073290 0.5073290 0.5073290
#>  [8,] 1.4791992 1.4791992 1.4791992
#>  [9,] 2.1445460 2.1445460 2.1445460
#> [10,] 1.5042314 1.5042314 1.5042314

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.2619405 0.9416306 2.8845400
#>  [2,] 0.7982606 0.8114185 0.7982606
#>  [3,] 0.6603016 3.3529139 0.2993758
#>  [4,] 2.2669001 2.7263083 3.2829751
#>  [5,] 1.8126382 3.6773230 1.9514189
#>  [6,] 0.1177819 1.0793267 0.1177819
#>  [7,] 2.1958390 2.1958390 0.8933234
#>  [8,] 0.7577786 0.2341108 2.1169605
#>  [9,] 2.2926930 0.2988579 2.2926930
#> [10,] 0.2488115 1.6568613 1.6568613
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 6.4933561 0.91913292 0.6619482
#>  [2,] 0.3259558 0.39369584 5.1206354
#>  [3,] 2.6645154 1.60520005 0.7063788
#>  [4,] 1.5008467 1.65014478 0.5384189
#>  [5,] 2.9609568 1.15487283 0.1191249
#>  [6,] 1.4608222 0.37525909 0.6004216
#>  [7,] 1.3044459 1.30275560 0.8329012
#>  [8,] 0.6852072 0.01178067 2.1705444
#>  [9,] 0.4351290 3.94268274 1.0745880
#> [10,] 1.0367231 2.22120438 0.6620735
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>               [,1]         [,2]         [,3]
#>  [1,] 1.9944586753 1.9944586753 1.9944586753
#>  [2,] 2.6785349118 2.6785349118 2.6785349118
#>  [3,] 0.5064911526 0.5064911526 0.5064911526
#>  [4,] 1.7661131377 1.7661131377 1.7661131377
#>  [5,] 0.0052699065 0.0052699065 0.0052699065
#>  [6,] 0.0007912693 0.0007912693 0.0007912693
#>  [7,] 3.0674602893 3.0674602893 3.0674602893
#>  [8,] 1.3105589794 1.3105589794 1.3105589794
#>  [9,] 1.5755596231 1.5755596231 1.5755596231
#> [10,] 1.4397641705 1.4397641705 1.4397641705