Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 1.27108418 1.95548827 4.15486600
#>  [2,] 0.80720285 0.98730940 0.56006176
#>  [3,] 0.85821871 0.86918592 0.86918592
#>  [4,] 0.28842862 0.54883299 0.54883299
#>  [5,] 1.20076635 1.97237318 2.33924497
#>  [6,] 0.10529846 0.10529846 0.10529846
#>  [7,] 0.08920267 1.06429918 0.15632547
#>  [8,] 1.34168612 0.30790141 0.30790141
#>  [9,] 1.29843612 0.80373267 0.62638732
#> [10,] 0.06305461 0.06305461 0.06305461
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33049174 1.61966029 0.72428363
#>  [2,] 0.11415819 1.47156059 0.73956587
#>  [3,] 0.56670955 2.15114205 0.49590843
#>  [4,] 0.66160653 2.78355583 1.28351096
#>  [5,] 0.04026561 0.08047342 0.86435207
#>  [6,] 0.53171948 1.18885087 0.08585557
#>  [7,] 0.65843385 0.14552267 1.50673922
#>  [8,] 0.53554261 2.93933672 0.77060574
#>  [9,] 0.70536902 0.95141710 1.61441271
#> [10,] 3.07652426 3.10904491 0.45810647
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1894654 1.1894654 1.1894654
#>  [2,] 0.1706705 0.1706705 0.1706705
#>  [3,] 0.7696476 0.7696476 0.7696476
#>  [4,] 0.2399645 0.2399645 0.2399645
#>  [5,] 0.7775073 0.7775073 0.7775073
#>  [6,] 0.3154498 0.3154498 0.3154498
#>  [7,] 2.4061871 2.4061871 2.4061871
#>  [8,] 1.4372050 1.4372050 1.4372050
#>  [9,] 1.1376471 1.1376471 1.1376471
#> [10,] 1.1384795 1.1384795 1.1384795

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3741557 0.6120569 1.3741557
#>  [2,] 0.4532952 0.4532952 0.2662130
#>  [3,] 1.1878238 0.2194027 3.8901832
#>  [4,] 0.7303452 0.7303452 0.7303452
#>  [5,] 2.3628898 2.3628898 2.3628898
#>  [6,] 0.7898121 0.1506413 0.1506413
#>  [7,] 0.8664019 0.6637791 0.4070879
#>  [8,] 0.4884117 0.4884117 0.4884117
#>  [9,] 3.7519064 0.7308438 0.4388167
#> [10,] 1.2324022 0.9145794 0.6480303
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.51400786 0.4962497 0.1009014
#>  [2,] 0.15211321 2.9564444 1.2056637
#>  [3,] 0.92329188 0.9354001 0.2092679
#>  [4,] 0.22586257 1.1848258 0.7872830
#>  [5,] 0.53152603 0.1185526 0.7409859
#>  [6,] 1.07511587 3.8003820 1.2112478
#>  [7,] 0.36697076 0.1536486 0.4620870
#>  [8,] 0.03295954 0.8480730 0.2679903
#>  [9,] 1.07610934 0.4307020 0.4013643
#> [10,] 0.95553282 0.2288816 0.4697775
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37738975 0.37738975 0.37738975
#>  [2,] 3.00275836 3.00275836 3.00275836
#>  [3,] 0.02572826 0.02572826 0.02572826
#>  [4,] 0.56931203 0.56931203 0.56931203
#>  [5,] 0.96845173 0.96845173 0.96845173
#>  [6,] 1.32990187 1.32990187 1.32990187
#>  [7,] 3.30952641 3.30952641 3.30952641
#>  [8,] 0.04715858 0.04715858 0.04715858
#>  [9,] 0.74777887 0.74777887 0.74777887
#> [10,] 1.55688612 1.55688612 1.55688612

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1287124 3.5782525 0.1287124
#>  [2,] 0.2955360 1.5119474 0.3362213
#>  [3,] 0.1319508 0.1319508 0.1058862
#>  [4,] 1.3012020 2.8183490 0.1949398
#>  [5,] 0.4169936 1.7187897 0.6577734
#>  [6,] 2.5226028 1.5217915 0.5461817
#>  [7,] 0.0593350 0.9048307 0.0593350
#>  [8,] 3.4747532 0.7315004 1.3668495
#>  [9,] 0.7543863 1.8007581 1.8007581
#> [10,] 0.2321985 0.4894328 0.4174466
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]        [,3]
#>  [1,] 1.7608244 0.80567896 2.651545828
#>  [2,] 1.2393870 0.87751204 2.732227632
#>  [3,] 0.3859137 0.30698682 0.244697290
#>  [4,] 0.5760839 1.04791493 1.767340560
#>  [5,] 0.7413091 0.35321763 0.307955722
#>  [6,] 2.6278803 0.04934854 0.004564139
#>  [7,] 0.5743127 0.06653378 0.332611777
#>  [8,] 1.2963282 1.24417616 1.030067749
#>  [9,] 0.1581094 0.15530238 0.230401052
#> [10,] 3.1935611 0.66138016 1.033287609
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8859160 0.8859160 0.8859160
#>  [2,] 0.1099239 0.1099239 0.1099239
#>  [3,] 1.0323600 1.0323600 1.0323600
#>  [4,] 2.4644244 2.4644244 2.4644244
#>  [5,] 3.5987897 3.5987897 3.5987897
#>  [6,] 0.3585229 0.3585229 0.3585229
#>  [7,] 0.5056523 0.5056523 0.5056523
#>  [8,] 1.7235858 1.7235858 1.7235858
#>  [9,] 1.2231227 1.2231227 1.2231227
#> [10,] 0.4181626 0.4181626 0.4181626