Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.41192060 2.73644079 0.16700944
#>  [2,] 2.90775541 2.91344353 2.43746134
#>  [3,] 0.07441419 0.07441419 0.07441419
#>  [4,] 0.30423063 0.30423063 0.53610006
#>  [5,] 0.71423943 2.18384876 2.18384876
#>  [6,] 1.35469404 1.01831020 0.36753060
#>  [7,] 2.26340486 0.63940935 0.75156341
#>  [8,] 1.35396074 1.35396074 1.35396074
#>  [9,] 2.29093197 1.03583425 2.83835068
#> [10,] 0.30646392 2.82391302 0.99312162
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.24512713 2.69046801 0.37991811
#>  [2,] 1.79072363 0.01782698 0.10206978
#>  [3,] 0.19988862 0.45089787 1.30785405
#>  [4,] 0.22133116 0.44567018 0.03368852
#>  [5,] 3.74852053 0.49332096 4.25190872
#>  [6,] 0.49048954 0.98950181 1.14290871
#>  [7,] 0.74134086 4.19174512 0.97987245
#>  [8,] 0.03315219 1.27229237 2.59803506
#>  [9,] 0.70327069 0.47940514 1.74355934
#> [10,] 0.57658070 0.37974338 2.11268157
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3620891 0.3620891 0.3620891
#>  [2,] 0.2438950 0.2438950 0.2438950
#>  [3,] 1.9473950 1.9473950 1.9473950
#>  [4,] 1.1430395 1.1430395 1.1430395
#>  [5,] 0.2615539 0.2615539 0.2615539
#>  [6,] 0.1764889 0.1764889 0.1764889
#>  [7,] 0.2226872 0.2226872 0.2226872
#>  [8,] 0.4572866 0.4572866 0.4572866
#>  [9,] 1.2028307 1.2028307 1.2028307
#> [10,] 0.4083755 0.4083755 0.4083755

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3004649 0.3004649 0.3004649
#>  [2,] 2.0490912 0.1648558 3.6484999
#>  [3,] 0.0116514 0.0116514 0.0116514
#>  [4,] 1.5143583 1.0736118 1.6467736
#>  [5,] 1.7645841 1.6297254 0.1727234
#>  [6,] 2.3737786 1.6795322 2.3737786
#>  [7,] 0.3042355 0.3042355 0.3042355
#>  [8,] 0.3668142 0.4011700 0.4011700
#>  [9,] 4.0525751 1.2850480 3.1196437
#> [10,] 0.6110862 0.6110862 0.6110862
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.04292404 1.3947615 1.4152695
#>  [2,] 0.90354597 2.0494596 0.5734730
#>  [3,] 1.27679428 0.6309676 0.3363520
#>  [4,] 0.58749121 0.2558323 1.1167332
#>  [5,] 0.94810038 1.4801734 2.3884915
#>  [6,] 0.41065826 0.6710895 2.1208545
#>  [7,] 0.43180487 0.1038609 0.3365928
#>  [8,] 0.40590253 0.2741850 1.1651527
#>  [9,] 0.18375532 0.4113577 0.1156663
#> [10,] 3.38491259 1.4908376 0.4559872
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.75139929 1.75139929 1.75139929
#>  [2,] 0.15573600 0.15573600 0.15573600
#>  [3,] 1.46994981 1.46994981 1.46994981
#>  [4,] 0.96545192 0.96545192 0.96545192
#>  [5,] 0.32164195 0.32164195 0.32164195
#>  [6,] 0.17161972 0.17161972 0.17161972
#>  [7,] 0.47532352 0.47532352 0.47532352
#>  [8,] 0.02360469 0.02360469 0.02360469
#>  [9,] 1.19210434 1.19210434 1.19210434
#> [10,] 0.65864484 0.65864484 0.65864484

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8572689 1.65426263 0.3779127
#>  [2,] 0.5849601 1.02841366 2.1548435
#>  [3,] 0.7839827 0.55788011 0.5578801
#>  [4,] 0.6170689 0.61706887 0.6170689
#>  [5,] 1.0877972 1.08779723 1.0877972
#>  [6,] 1.1282268 0.56945502 0.5694550
#>  [7,] 0.3725159 0.37251590 0.3725159
#>  [8,] 0.1459310 0.04692056 0.1459310
#>  [9,] 1.6347443 3.46760570 2.2369277
#> [10,] 0.1419714 0.32135931 0.0301366
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.13961221 0.9593964 0.12982679
#>  [2,] 0.96131890 0.1525618 0.03681591
#>  [3,] 0.25794222 0.2055903 1.46621332
#>  [4,] 0.01682429 0.4483838 0.34816405
#>  [5,] 0.87815061 1.4310339 0.52265881
#>  [6,] 0.53219525 0.8562624 0.29947438
#>  [7,] 0.45204079 0.4763311 0.07962337
#>  [8,] 1.27071519 0.9853702 2.61917041
#>  [9,] 0.85482036 0.1325445 0.51908082
#> [10,] 0.67456396 0.3711140 0.22879384
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.88084075 1.88084075 1.88084075
#>  [2,] 0.21079706 0.21079706 0.21079706
#>  [3,] 0.74498389 0.74498389 0.74498389
#>  [4,] 0.37522369 0.37522369 0.37522369
#>  [5,] 0.11972921 0.11972921 0.11972921
#>  [6,] 0.19734109 0.19734109 0.19734109
#>  [7,] 1.72351666 1.72351666 1.72351666
#>  [8,] 2.49863636 2.49863636 2.49863636
#>  [9,] 0.04691822 0.04691822 0.04691822
#> [10,] 1.15689892 1.15689892 1.15689892