Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]      [,2]       [,3]
#>  [1,] 0.12961752 1.8115318 0.13155465
#>  [2,] 0.02877775 0.8227306 0.05493398
#>  [3,] 0.27380760 0.7396560 2.79219795
#>  [4,] 1.10190512 1.1019051 1.10190512
#>  [5,] 0.21324551 0.2132455 0.21324551
#>  [6,] 0.51201574 0.5120157 0.05016929
#>  [7,] 0.10611306 0.8305592 0.10611306
#>  [8,] 0.41025038 0.4102504 0.41025038
#>  [9,] 0.35109344 0.3510934 0.35109344
#> [10,] 0.51456776 0.5145678 0.51456776
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]        [,2]       [,3]
#>  [1,] 0.88039422 0.736886257 0.01213735
#>  [2,] 0.78262988 0.678324044 4.37411001
#>  [3,] 0.12100373 0.206700511 0.52252141
#>  [4,] 0.08359885 0.490803940 1.84266836
#>  [5,] 1.69871398 0.122598766 0.69806514
#>  [6,] 1.50786550 0.060269907 1.57376267
#>  [7,] 0.82781420 1.717500063 0.46084702
#>  [8,] 2.13881471 0.476364252 3.35011751
#>  [9,] 0.65236555 0.007927091 2.00025141
#> [10,] 0.24705331 2.360974803 0.01666174
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7926815 0.7926815 0.7926815
#>  [2,] 1.1704449 1.1704449 1.1704449
#>  [3,] 3.2816360 3.2816360 3.2816360
#>  [4,] 1.6397119 1.6397119 1.6397119
#>  [5,] 0.2914812 0.2914812 0.2914812
#>  [6,] 1.2108310 1.2108310 1.2108310
#>  [7,] 0.8206919 0.8206919 0.8206919
#>  [8,] 0.1335343 0.1335343 0.1335343
#>  [9,] 0.7251857 0.7251857 0.7251857
#> [10,] 1.0253489 1.0253489 1.0253489

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.15062713 2.9159611 0.15062713
#>  [2,] 0.35645257 1.3661654 1.36616537
#>  [3,] 0.07089798 0.7275334 1.97335612
#>  [4,] 0.46873671 0.4687367 0.46873671
#>  [5,] 0.34028301 0.2210496 0.22104964
#>  [6,] 0.14838288 0.1483829 0.14838288
#>  [7,] 0.28822967 5.7079121 0.38730108
#>  [8,] 0.21316045 0.2131604 0.21316045
#>  [9,] 0.49085170 0.4908517 0.49085170
#> [10,] 1.26027705 1.6751333 0.09709118
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 2.73482067 0.90016220 0.5538070
#>  [2,] 0.02797548 0.65805531 2.2148828
#>  [3,] 0.15142238 0.08687698 2.1850692
#>  [4,] 2.07222742 0.92195320 1.2951512
#>  [5,] 2.39954526 1.91342365 0.4911562
#>  [6,] 0.07349370 0.29578569 0.1939100
#>  [7,] 0.57456113 0.31187814 0.3300529
#>  [8,] 0.56378300 1.57404709 1.6353355
#>  [9,] 0.02352567 1.32811772 1.3820055
#> [10,] 2.51517139 1.30397199 0.2451131
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55522573 0.55522573 0.55522573
#>  [2,] 0.04860324 0.04860324 0.04860324
#>  [3,] 1.04665808 1.04665808 1.04665808
#>  [4,] 0.02529968 0.02529968 0.02529968
#>  [5,] 1.08623704 1.08623704 1.08623704
#>  [6,] 0.29294746 0.29294746 0.29294746
#>  [7,] 1.47270229 1.47270229 1.47270229
#>  [8,] 1.84363666 1.84363666 1.84363666
#>  [9,] 0.01970505 0.01970505 0.01970505
#> [10,] 2.44496771 2.44496771 2.44496771

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7260124 3.1432288 1.2846240
#>  [2,] 1.9641923 1.2448307 1.1630244
#>  [3,] 0.7426887 0.7426887 1.5008059
#>  [4,] 0.2377975 1.7111736 1.7111736
#>  [5,] 0.1611906 0.1611906 0.1611906
#>  [6,] 1.7133404 1.0660750 2.1820224
#>  [7,] 1.6585433 0.8487724 1.6585433
#>  [8,] 2.3070655 2.3070655 0.3870457
#>  [9,] 0.3306071 0.3306071 0.3306071
#> [10,] 0.4629738 0.4629738 0.4629738
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.22543275 0.86634087 1.5767801
#>  [2,] 0.08036525 0.69304553 0.1302459
#>  [3,] 1.08965699 1.08219300 0.1033909
#>  [4,] 1.07443114 0.09398771 6.4653450
#>  [5,] 1.32840971 0.12500022 0.1787800
#>  [6,] 1.06917949 0.33398643 0.2943207
#>  [7,] 0.65427322 0.17012532 0.7944848
#>  [8,] 0.30006014 0.41087241 0.3512806
#>  [9,] 2.41460582 0.30080716 1.7378016
#> [10,] 0.74623133 1.79530664 0.8651081
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0814130 2.0814130 2.0814130
#>  [2,] 1.3469928 1.3469928 1.3469928
#>  [3,] 0.3108537 0.3108537 0.3108537
#>  [4,] 0.8014318 0.8014318 0.8014318
#>  [5,] 0.6039251 0.6039251 0.6039251
#>  [6,] 0.7936640 0.7936640 0.7936640
#>  [7,] 0.4565770 0.4565770 0.4565770
#>  [8,] 0.2289420 0.2289420 0.2289420
#>  [9,] 1.9841071 1.9841071 1.9841071
#> [10,] 0.8394441 0.8394441 0.8394441