Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.28740493 0.6781963 0.28740493
#>  [2,] 1.13131500 1.1313150 1.13131500
#>  [3,] 2.86261768 0.6111656 0.38521134
#>  [4,] 0.39484696 0.3696326 0.39484696
#>  [5,] 1.37353480 0.6012055 0.91386285
#>  [6,] 0.82647234 0.8264723 0.82647234
#>  [7,] 0.04977085 1.7434815 0.04977085
#>  [8,] 0.12710562 0.1271056 0.12710562
#>  [9,] 0.71769022 2.1714095 0.28656769
#> [10,] 0.67422714 1.5188161 1.52599653
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.11437551 1.6841853 0.98174158
#>  [2,] 1.10059212 1.2578753 0.47488787
#>  [3,] 1.64147731 1.3204116 1.11537489
#>  [4,] 3.88478280 3.2374226 0.93068317
#>  [5,] 0.49397720 1.6403267 0.06964429
#>  [6,] 0.18141132 1.6306811 0.37823526
#>  [7,] 3.69832419 1.3701188 0.41232692
#>  [8,] 0.15566791 0.1689246 0.09159842
#>  [9,] 0.74810827 1.2161793 0.19430165
#> [10,] 0.09469741 4.6849555 2.65278464
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14511446 0.14511446 0.14511446
#>  [2,] 0.13144014 0.13144014 0.13144014
#>  [3,] 0.01475437 0.01475437 0.01475437
#>  [4,] 0.18465715 0.18465715 0.18465715
#>  [5,] 0.10918550 0.10918550 0.10918550
#>  [6,] 1.46769604 1.46769604 1.46769604
#>  [7,] 0.39703404 0.39703404 0.39703404
#>  [8,] 3.59569631 3.59569631 3.59569631
#>  [9,] 0.11950745 0.11950745 0.11950745
#> [10,] 0.91892256 0.91892256 0.91892256

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>            [,1]        [,2]      [,3]
#>  [1,] 0.7087980 0.003272865 1.0384634
#>  [2,] 0.5782077 0.366180762 0.5782077
#>  [3,] 7.2027583 0.626728766 0.6267288
#>  [4,] 0.3745028 1.214993000 0.6645101
#>  [5,] 1.2929375 0.359558982 1.7497090
#>  [6,] 0.7181199 1.622341457 0.9284852
#>  [7,] 0.2770416 0.277041594 0.2770416
#>  [8,] 0.2188216 0.218821582 0.2188216
#>  [9,] 3.7422379 3.049783562 0.7365971
#> [10,] 1.7439421 1.743942065 0.5927524
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.02259058 0.3097636 0.8803592
#>  [2,] 0.24600167 0.5680755 0.8403890
#>  [3,] 3.34554645 0.4210152 0.4004895
#>  [4,] 0.15268478 2.5271969 0.4034751
#>  [5,] 0.54564693 0.7761090 0.1933290
#>  [6,] 0.52033154 4.1033941 0.2248464
#>  [7,] 0.48883695 1.5356036 1.2786622
#>  [8,] 0.16691247 0.4853396 0.2002699
#>  [9,] 2.53162508 0.8164150 0.6681655
#> [10,] 0.19052376 0.8244276 0.2215350
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.95361582 0.95361582 0.95361582
#>  [2,] 0.11578118 0.11578118 0.11578118
#>  [3,] 0.05009133 0.05009133 0.05009133
#>  [4,] 0.61498145 0.61498145 0.61498145
#>  [5,] 1.22136232 1.22136232 1.22136232
#>  [6,] 0.68307329 0.68307329 0.68307329
#>  [7,] 0.60707265 0.60707265 0.60707265
#>  [8,] 2.83210637 2.83210637 2.83210637
#>  [9,] 2.82745795 2.82745795 2.82745795
#> [10,] 1.80700037 1.80700037 1.80700037

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.74865542 3.5025677 0.2190191
#>  [2,] 0.54224174 0.6921723 0.2280220
#>  [3,] 2.06389064 2.0638906 0.2027146
#>  [4,] 1.42142018 0.4667885 1.0226036
#>  [5,] 2.38308584 0.7482139 0.4512730
#>  [6,] 1.23300836 1.8853433 1.5055992
#>  [7,] 1.56787786 1.5678779 1.5678779
#>  [8,] 0.05273283 0.4358783 0.1947068
#>  [9,] 0.97120635 1.1369629 1.1369629
#> [10,] 0.77878136 0.7449433 0.3379010
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.62412416 0.8124904 1.0673845
#>  [2,] 0.02159569 0.4929369 0.4990852
#>  [3,] 3.19324400 0.5652868 0.8761788
#>  [4,] 0.64189719 0.4749411 0.7583075
#>  [5,] 0.13496330 0.5925183 0.1336505
#>  [6,] 1.05526118 1.5656104 0.3722389
#>  [7,] 0.92997262 1.3833968 1.0606180
#>  [8,] 0.45912670 2.1666614 0.9485421
#>  [9,] 0.98535004 0.3647245 2.5032633
#> [10,] 1.13919250 1.8324268 4.3443576
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9792486 0.9792486 0.9792486
#>  [2,] 0.5902436 0.5902436 0.5902436
#>  [3,] 0.6328610 0.6328610 0.6328610
#>  [4,] 1.5780038 1.5780038 1.5780038
#>  [5,] 0.4315677 0.4315677 0.4315677
#>  [6,] 2.4344729 2.4344729 2.4344729
#>  [7,] 0.0686649 0.0686649 0.0686649
#>  [8,] 2.9475529 2.9475529 2.9475529
#>  [9,] 0.1396971 0.1396971 0.1396971
#> [10,] 0.1303472 0.1303472 0.1303472