Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3840782 0.7232955 1.2011782
#>  [2,] 0.8179462 1.2443563 1.2443563
#>  [3,] 0.3258234 0.3258234 0.3258234
#>  [4,] 0.5871338 0.5871338 1.2673632
#>  [5,] 1.6552913 0.2658851 1.6552913
#>  [6,] 0.1867060 0.1549361 0.5448318
#>  [7,] 0.3343416 0.6182227 0.6182227
#>  [8,] 0.5458849 0.8905787 0.4936077
#>  [9,] 0.8498496 2.0233784 0.6036075
#> [10,] 0.4232763 0.4232763 0.4470367
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.59164387 0.0185294 1.1212152
#>  [2,] 0.29883369 0.7243059 1.2261396
#>  [3,] 0.09574859 1.3815999 1.4075060
#>  [4,] 5.54517086 0.9047455 1.0212953
#>  [5,] 1.09233710 0.1139241 0.5666601
#>  [6,] 0.43619508 0.9910618 1.2745586
#>  [7,] 0.10860737 0.3053809 3.8322682
#>  [8,] 0.95993851 0.4436652 1.7953410
#>  [9,] 0.80601449 0.3767674 0.9535505
#> [10,] 0.25836955 4.4795369 1.4406107
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7267491 0.7267491 0.7267491
#>  [2,] 0.2657503 0.2657503 0.2657503
#>  [3,] 0.1886026 0.1886026 0.1886026
#>  [4,] 4.0090383 4.0090383 4.0090383
#>  [5,] 0.2290448 0.2290448 0.2290448
#>  [6,] 2.6042037 2.6042037 2.6042037
#>  [7,] 0.1500083 0.1500083 0.1500083
#>  [8,] 0.7891430 0.7891430 0.7891430
#>  [9,] 0.6568265 0.6568265 0.6568265
#> [10,] 0.2717797 0.2717797 0.2717797

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58301407 0.58301407 0.58301407
#>  [2,] 1.29278059 1.29278059 0.03131260
#>  [3,] 1.09120339 1.90826602 0.44609594
#>  [4,] 0.87258059 1.53239381 3.58897167
#>  [5,] 0.16837850 0.55296054 1.44165524
#>  [6,] 2.86723897 0.07041280 2.86723897
#>  [7,] 1.59307662 0.01115831 0.08211927
#>  [8,] 0.71868753 2.34001274 0.85007971
#>  [9,] 0.04854873 0.74419256 0.53507252
#> [10,] 0.87020864 0.23678437 0.23678437
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5005273 0.09349263 1.0899523
#>  [2,] 0.1283282 0.10972466 0.3012010
#>  [3,] 0.9211535 2.96857195 2.0981512
#>  [4,] 0.0442376 0.05013454 0.3115392
#>  [5,] 0.4829006 0.37749441 0.8552303
#>  [6,] 3.1079451 0.27348949 0.6404629
#>  [7,] 0.0433910 0.21248135 2.2606123
#>  [8,] 1.3321367 0.28712297 4.2945013
#>  [9,] 0.5141529 0.06220983 1.8142612
#> [10,] 0.7651714 0.86228048 0.2569374
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07922239 1.07922239 1.07922239
#>  [2,] 2.10204490 2.10204490 2.10204490
#>  [3,] 0.94017003 0.94017003 0.94017003
#>  [4,] 0.03040106 0.03040106 0.03040106
#>  [5,] 1.55067949 1.55067949 1.55067949
#>  [6,] 0.59018004 0.59018004 0.59018004
#>  [7,] 0.99305203 0.99305203 0.99305203
#>  [8,] 0.54168888 0.54168888 0.54168888
#>  [9,] 1.02555929 1.02555929 1.02555929
#> [10,] 1.81733426 1.81733426 1.81733426

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1405991 1.1854522 1.1854522
#>  [2,] 0.3497592 0.3497592 0.1668224
#>  [3,] 0.9648875 0.9648875 0.9648875
#>  [4,] 0.2335400 0.2335400 0.2335400
#>  [5,] 0.2270108 1.5901568 1.5901568
#>  [6,] 0.9206474 1.3045512 1.3045512
#>  [7,] 1.6402310 1.6402310 1.6402310
#>  [8,] 0.6894707 0.6894707 0.8193954
#>  [9,] 2.4640997 1.6974521 0.4509724
#> [10,] 0.5736079 3.6209463 5.7836567
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35522680 1.07952511 0.57328775
#>  [2,] 0.06428333 1.13139452 0.06593515
#>  [3,] 1.27138074 0.39542933 0.27055646
#>  [4,] 0.28836349 0.47092866 1.64319021
#>  [5,] 1.52717630 0.02888775 0.28433142
#>  [6,] 1.74780447 0.10149063 0.29147076
#>  [7,] 0.77793445 1.83644390 1.28492898
#>  [8,] 4.08299787 1.15898573 0.21026427
#>  [9,] 0.44323109 0.13465346 1.46996100
#> [10,] 2.14893723 0.12853932 0.53263843
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.74248237 1.74248237 1.74248237
#>  [2,] 1.35975428 1.35975428 1.35975428
#>  [3,] 0.48739016 0.48739016 0.48739016
#>  [4,] 0.09748002 0.09748002 0.09748002
#>  [5,] 0.06056247 0.06056247 0.06056247
#>  [6,] 2.70336410 2.70336410 2.70336410
#>  [7,] 1.52772906 1.52772906 1.52772906
#>  [8,] 0.52027571 0.52027571 0.52027571
#>  [9,] 0.10717163 0.10717163 0.10717163
#> [10,] 0.71747262 0.71747262 0.71747262