Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]      [,3]
#>  [1,] 0.98209140 0.98209140 1.5602680
#>  [2,] 1.69381206 0.31494134 1.6938121
#>  [3,] 0.92741024 0.92741024 0.9274102
#>  [4,] 0.73378508 0.73378508 0.7337851
#>  [5,] 0.03764050 1.06040795 0.0376405
#>  [6,] 0.61037000 0.23376774 0.1556830
#>  [7,] 0.64125419 0.03139258 0.6412542
#>  [8,] 0.08763124 1.13098086 1.1309809
#>  [9,] 0.55777946 0.60686336 0.8850441
#> [10,] 2.22596922 0.34886586 0.3488659
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.66551140 0.4581235 0.20198192
#>  [2,] 1.00509675 0.4699794 0.89214556
#>  [3,] 1.22573173 0.9044999 3.16953627
#>  [4,] 2.53383728 0.1082710 1.00553531
#>  [5,] 0.04220633 1.1709897 0.10544601
#>  [6,] 1.03010057 0.6043718 1.33438066
#>  [7,] 1.14267901 1.3812944 1.85723001
#>  [8,] 0.84804394 0.7970001 0.93353777
#>  [9,] 0.61886540 4.1293263 0.06417228
#> [10,] 0.70498645 2.1188649 0.04830591
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9181241 1.9181241 1.9181241
#>  [2,] 0.2217292 0.2217292 0.2217292
#>  [3,] 0.0668080 0.0668080 0.0668080
#>  [4,] 0.2402531 0.2402531 0.2402531
#>  [5,] 1.5362461 1.5362461 1.5362461
#>  [6,] 0.8330725 0.8330725 0.8330725
#>  [7,] 0.4978065 0.4978065 0.4978065
#>  [8,] 2.3195795 2.3195795 2.3195795
#>  [9,] 0.4245900 0.4245900 0.4245900
#> [10,] 0.1876175 0.1876175 0.1876175

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9347887 2.35422344 1.7204467
#>  [2,] 0.7325755 1.06765136 0.7325755
#>  [3,] 0.7064525 1.26799134 1.2679913
#>  [4,] 0.1307353 0.01173062 0.1307353
#>  [5,] 1.0037193 1.00371932 1.0037193
#>  [6,] 2.3378188 0.32641802 0.7282135
#>  [7,] 1.5581019 1.55810187 0.6401446
#>  [8,] 0.3217914 3.71498742 3.1173824
#>  [9,] 0.4658106 0.46581055 1.5982102
#> [10,] 0.6286535 0.28851806 0.6056565
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3483373 1.41547964 2.70997467
#>  [2,] 1.5567816 0.03797891 0.36938431
#>  [3,] 1.2788295 1.37018086 0.58723288
#>  [4,] 0.7608747 0.86776222 2.17969581
#>  [5,] 4.1793679 1.18985596 3.69214544
#>  [6,] 2.1240535 1.88780493 0.05980918
#>  [7,] 1.6210828 0.83790534 0.48436248
#>  [8,] 2.8589831 1.75208659 0.58187985
#>  [9,] 1.4916946 0.84061844 0.78982992
#> [10,] 0.6952584 1.40849818 1.36069204
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.0665584 0.0665584 0.0665584
#>  [2,] 0.6513265 0.6513265 0.6513265
#>  [3,] 0.9315011 0.9315011 0.9315011
#>  [4,] 0.1324277 0.1324277 0.1324277
#>  [5,] 0.1598685 0.1598685 0.1598685
#>  [6,] 0.4099893 0.4099893 0.4099893
#>  [7,] 1.4721458 1.4721458 1.4721458
#>  [8,] 0.9496527 0.9496527 0.9496527
#>  [9,] 1.4386576 1.4386576 1.4386576
#> [10,] 4.1854830 4.1854830 4.1854830

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.86978790 0.8697879 0.5470714
#>  [2,] 0.20952532 0.2106437 0.2095253
#>  [3,] 0.18962230 0.2984304 0.3806194
#>  [4,] 0.06534228 0.1206059 0.4478397
#>  [5,] 0.56352559 0.5635256 0.5635256
#>  [6,] 1.21621753 0.6260686 0.3564201
#>  [7,] 0.85775301 0.1945333 1.4938804
#>  [8,] 1.12443925 1.2317501 0.5275226
#>  [9,] 0.11205263 0.1720786 0.4120334
#> [10,] 0.16272072 0.1627207 2.6288822
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3106608 0.69284663 0.11440200
#>  [2,] 0.9872907 0.73898557 2.20730870
#>  [3,] 3.2628376 1.47029251 0.67430434
#>  [4,] 4.0159806 6.54551729 1.21016907
#>  [5,] 0.1680836 0.07019955 0.58143386
#>  [6,] 0.7413525 1.87959424 0.02862601
#>  [7,] 0.8456537 0.18077336 0.31231705
#>  [8,] 0.5137530 1.30670407 0.12678194
#>  [9,] 0.1756390 1.24614511 0.18588353
#> [10,] 1.6529970 2.15654148 2.84088777
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6037237 0.6037237 0.6037237
#>  [2,] 0.6332227 0.6332227 0.6332227
#>  [3,] 0.2327994 0.2327994 0.2327994
#>  [4,] 1.4483820 1.4483820 1.4483820
#>  [5,] 0.1294947 0.1294947 0.1294947
#>  [6,] 0.5439727 0.5439727 0.5439727
#>  [7,] 0.5405113 0.5405113 0.5405113
#>  [8,] 0.3833901 0.3833901 0.3833901
#>  [9,] 1.0073679 1.0073679 1.0073679
#> [10,] 0.4110160 0.4110160 0.4110160