Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.85599821 0.60713323 0.46664998
#>  [2,] 0.43786189 1.92253956 0.57096665
#>  [3,] 0.73860867 0.88224807 0.37845507
#>  [4,] 0.11490206 0.11490206 0.11490206
#>  [5,] 0.25539700 0.01664677 0.25539700
#>  [6,] 0.03488559 0.94837761 0.03488559
#>  [7,] 0.52273490 0.81236295 1.91840705
#>  [8,] 1.18090830 0.67865018 1.18090830
#>  [9,] 0.18508980 0.18508980 0.18508980
#> [10,] 1.10043466 1.10043466 1.10043466
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0))
#>             [,1]       [,2]        [,3]
#>  [1,] 3.17647505 0.27899154 0.512776904
#>  [2,] 0.58338161 0.05557640 0.008636903
#>  [3,] 0.90006780 0.32209270 0.515382351
#>  [4,] 0.62309604 1.23305549 2.067225652
#>  [5,] 0.23411716 0.07978611 0.133336305
#>  [6,] 0.39628815 1.61762830 0.084616885
#>  [7,] 0.14718642 1.20913629 1.047590920
#>  [8,] 2.56961613 1.28521776 1.610533591
#>  [9,] 0.10507249 0.39621051 0.523877251
#> [10,] 0.04662244 0.23920355 2.957708040
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47216784 0.47216784 0.47216784
#>  [2,] 1.02862117 1.02862117 1.02862117
#>  [3,] 5.13589351 5.13589351 5.13589351
#>  [4,] 2.74649437 2.74649437 2.74649437
#>  [5,] 1.98795010 1.98795010 1.98795010
#>  [6,] 2.07985623 2.07985623 2.07985623
#>  [7,] 1.58423258 1.58423258 1.58423258
#>  [8,] 0.49131977 0.49131977 0.49131977
#>  [9,] 0.55766260 0.55766260 0.55766260
#> [10,] 0.09118018 0.09118018 0.09118018

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6720451 1.27167894 0.4404264
#>  [2,] 0.5632974 0.25212575 0.5632974
#>  [3,] 0.1049502 0.10495018 0.1049502
#>  [4,] 0.1164659 0.11646590 0.1585430
#>  [5,] 1.0429591 4.14636799 0.8350632
#>  [6,] 0.7374653 0.05363885 0.7374653
#>  [7,] 0.9208550 0.92085497 0.4491817
#>  [8,] 0.1898628 0.18986280 0.1898628
#>  [9,] 0.1704127 0.17041275 0.1704127
#> [10,] 1.1404200 1.14041998 1.1404200
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 3.86754299 0.2604978 0.70843149
#>  [2,] 3.81787779 0.1310143 0.21839594
#>  [3,] 0.08632079 1.3068115 0.49260640
#>  [4,] 0.31426248 0.4220398 0.02192578
#>  [5,] 1.44026395 1.9820931 1.08817042
#>  [6,] 0.58072707 0.6232206 0.76178840
#>  [7,] 0.86173222 0.7527682 1.18218993
#>  [8,] 2.13627562 1.9692595 1.16481000
#>  [9,] 2.43676971 0.1657154 3.85643724
#> [10,] 0.01631555 0.4001585 1.25899129
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.351255486 0.351255486 0.351255486
#>  [2,] 1.652161723 1.652161723 1.652161723
#>  [3,] 1.229539085 1.229539085 1.229539085
#>  [4,] 0.882137954 0.882137954 0.882137954
#>  [5,] 0.505072676 0.505072676 0.505072676
#>  [6,] 1.415176425 1.415176425 1.415176425
#>  [7,] 0.929906556 0.929906556 0.929906556
#>  [8,] 0.003018857 0.003018857 0.003018857
#>  [9,] 2.939958198 2.939958198 2.939958198
#> [10,] 0.656843287 0.656843287 0.656843287

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33382895 0.32407483 0.51476836
#>  [2,] 0.06915356 0.06915356 0.06915356
#>  [3,] 0.50217526 1.49931580 1.49931580
#>  [4,] 0.36299845 0.19580194 0.36299845
#>  [5,] 1.69841013 2.30319265 1.69841013
#>  [6,] 1.86453456 1.21801320 0.12261132
#>  [7,] 1.13356782 1.01792493 0.43025545
#>  [8,] 0.86634661 1.70562764 0.22707804
#>  [9,] 1.54151450 1.57284629 1.54151450
#> [10,] 0.70623806 0.70623806 0.70623806
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.15478937 0.002805254 0.89985788
#>  [2,] 0.08299739 0.562872375 2.23760247
#>  [3,] 0.23586071 0.266309572 1.14696902
#>  [4,] 0.04129820 0.504126267 0.62181804
#>  [5,] 0.32466224 1.333937338 0.70928804
#>  [6,] 0.71049136 1.625629074 0.08330783
#>  [7,] 3.74763411 0.251306020 0.37864140
#>  [8,] 0.54643537 0.130079138 0.95220639
#>  [9,] 1.71877042 1.795305261 0.07690737
#> [10,] 0.26400099 0.582844895 0.90869971
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.37824350 1.37824350 1.37824350
#>  [2,] 0.63929955 0.63929955 0.63929955
#>  [3,] 1.91889437 1.91889437 1.91889437
#>  [4,] 0.43627658 0.43627658 0.43627658
#>  [5,] 1.80478637 1.80478637 1.80478637
#>  [6,] 0.05780930 0.05780930 0.05780930
#>  [7,] 0.52167320 0.52167320 0.52167320
#>  [8,] 1.52142987 1.52142987 1.52142987
#>  [9,] 0.21568246 0.21568246 0.21568246
#> [10,] 0.02403388 0.02403388 0.02403388