Sample from Marshall–Olkin distributions

Draws n iid samples from a d-variate Marshall–Olkin distribution parametrized by shock-arrival intensities.

Usage

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock-arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM for small dimensions only; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with n rows and d columns, with the rows corresponding to iid distributed samples of a d-variate Marshall–Olkin distribution with shock-arrival intensities

intensities.

Details

The Marshall–Olkin distribution was introduced in (Marshall and Olkin 1967) and has the survival function $$ \bar{F}{(t)} = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t = {(t_{1}, \ldots, t_{d})} > 0 , $$ for shock-arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\). They are called shock-arrival intensities as they correspond to the rates of independent exponential random variables from the exogenous shock model (ESM), and a shock-arrival intensity \(\lambda_{I}\) of shock \(I\) equal to zero implies that the the shock \(I\) never arrives. We use the following binary representation to map a subsets of \(\{ 1, \ldots, d\}\) to integers \(0, \ldots, 2^d-1\): $$ I \equiv \sum_{k \in I}{ 2^{k-1} } $$

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via independent exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process with set-valued marks. The process is stopped when all components are hit by a shock, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi:10.2307/2282907 .

See also

Examples

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4)
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4332742 0.4332742 0.43327422
#>  [2,] 0.8145682 0.8145682 0.03151031
#>  [3,] 1.3708320 0.9047787 0.09326529
#>  [4,] 1.5954745 2.0711889 1.27717877
#>  [5,] 0.5294766 1.1230009 2.10474860
#>  [6,] 0.9305823 0.9305823 0.93058229
#>  [7,] 0.8974076 0.8750919 0.28896881
#>  [8,] 1.5391260 0.5022717 0.61891647
#>  [9,] 2.8814951 3.0170528 2.30618906
#> [10,] 3.6960035 1.4829399 1.48293988
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0)
)
#>               [,1]      [,2]        [,3]
#>  [1,] 0.3940955854 0.8706064 1.943909551
#>  [2,] 0.3551947841 0.7469686 0.569905644
#>  [3,] 0.0227393300 0.4465233 0.960865749
#>  [4,] 0.0002102099 1.4834823 0.175756043
#>  [5,] 0.4040795700 1.1655414 0.008227351
#>  [6,] 0.0348230400 0.5135286 2.281326488
#>  [7,] 0.8822388014 0.1833615 0.171067095
#>  [8,] 1.1716722869 1.3488559 0.242194423
#>  [9,] 1.8042933397 0.9954350 3.220663328
#> [10,] 2.6693951032 0.1663279 0.353476864
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1)
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.46651322 2.46651322 2.46651322
#>  [2,] 0.89715322 0.89715322 0.89715322
#>  [3,] 1.43826877 1.43826877 1.43826877
#>  [4,] 1.05912797 1.05912797 1.05912797
#>  [5,] 1.05011355 1.05011355 1.05011355
#>  [6,] 0.06073726 0.06073726 0.06073726
#>  [7,] 0.65751746 0.65751746 0.65751746
#>  [8,] 0.90476606 0.90476606 0.90476606
#>  [9,] 0.37999808 0.37999808 0.37999808
#> [10,] 0.04271985 0.04271985 0.04271985

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.36403652 0.3640365 0.3640365
#>  [2,] 1.17153968 1.1715397 0.1079069
#>  [3,] 0.74312307 0.7431231 0.7431231
#>  [4,] 1.73975039 1.7397504 0.8253085
#>  [5,] 0.82973884 1.6395737 0.8297388
#>  [6,] 0.43591398 0.4359140 0.4393783
#>  [7,] 0.62484205 0.6248421 0.6248421
#>  [8,] 0.16194916 0.2326689 0.1910355
#>  [9,] 0.09924827 0.1691088 0.4886671
#> [10,] 0.05703699 0.4611600 0.8080590
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.49064733 2.09771913 1.04907789
#>  [2,] 1.81822691 0.23556594 0.22907413
#>  [3,] 0.09344653 1.50445518 0.23949865
#>  [4,] 0.14397670 2.11503888 1.78652354
#>  [5,] 0.88815571 0.05440357 1.50834466
#>  [6,] 0.43935061 0.31626810 1.23239670
#>  [7,] 1.69180945 0.71257071 0.35546023
#>  [8,] 1.95754400 1.44980712 0.92792036
#>  [9,] 0.59500888 1.63187312 0.04804564
#> [10,] 0.06357031 4.74919524 0.96347019
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9268026 0.9268026 0.9268026
#>  [2,] 0.3334049 0.3334049 0.3334049
#>  [3,] 0.4428489 0.4428489 0.4428489
#>  [4,] 0.9256085 0.9256085 0.9256085
#>  [5,] 3.3959962 3.3959962 3.3959962
#>  [6,] 0.4600348 0.4600348 0.4600348
#>  [7,] 1.2979586 1.2979586 1.2979586
#>  [8,] 0.2105053 0.2105053 0.2105053
#>  [9,] 0.9708400 0.9708400 0.9708400
#> [10,] 0.1913857 0.1913857 0.1913857

rmo(
    10, 3,
    c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4),
    method = "AM"
)
#>            [,1]      [,2]        [,3]
#>  [1,] 1.9436973 1.5482172 4.026515820
#>  [2,] 0.4050212 0.4734998 0.988773304
#>  [3,] 0.0198282 0.4793550 0.019828196
#>  [4,] 1.1095575 1.6624793 0.636615692
#>  [5,] 0.6582095 0.4138662 0.658209526
#>  [6,] 0.7988457 0.7988457 0.798845695
#>  [7,] 0.2429575 0.2429575 0.242957458
#>  [8,] 0.8883830 0.5831010 3.226168074
#>  [9,] 0.5390355 0.5390355 0.539035515
#> [10,] 3.2653003 0.4069142 0.004700855
## independence
rmo(
    10, 3,
    c(1, 1, 0, 1, 0, 0, 0),
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1271035 0.29159540 0.68140616
#>  [2,] 2.3349345 1.02202658 0.71147489
#>  [3,] 1.7903798 0.22813933 0.01994426
#>  [4,] 1.5382849 0.49916848 0.72276661
#>  [5,] 1.4848845 0.12632143 0.13937087
#>  [6,] 0.6931330 0.41141111 1.71973152
#>  [7,] 0.5773699 1.68248076 0.35389091
#>  [8,] 0.7079033 0.07832118 1.89507670
#>  [9,] 0.4192670 0.45223916 0.76259686
#> [10,] 1.3253207 1.43375033 0.08914120
## comonotone
rmo(
    10, 3,
    c(0, 0, 0, 0, 0, 0, 1),
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3962076 1.3962076 1.3962076
#>  [2,] 1.2536454 1.2536454 1.2536454
#>  [3,] 0.4953417 0.4953417 0.4953417
#>  [4,] 0.8112017 0.8112017 0.8112017
#>  [5,] 0.4274337 0.4274337 0.4274337
#>  [6,] 0.3829850 0.3829850 0.3829850
#>  [7,] 5.4043767 5.4043767 5.4043767
#>  [8,] 1.0980798 1.0980798 1.0980798
#>  [9,] 0.1841928 0.1841928 0.1841928
#> [10,] 0.2952143 0.2952143 0.2952143