Sample from a Marshall–Olkin distribution

Draws \(n\) iid samples from a \(d\)-variate Marshall–Olkin distribution parametrized by shock arrival intensities.

rmo(n, d, intensities, method = c("AM", "ESM"))

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

intensities

a numeric vector for the shock arrival intensities.

method

a string indicating which sampling algorithm should be used. Use "AM" for the Arnold model and "ESM" for the exogenous shock model. We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate Marshall–Olkin distributionwith shock arrival intensitiesintensities.

Parametrization

The Marshall–Olkin distribution has the survival function $$ \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock arrival intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

Shock arrival intensities

The shock arrival intensities must be stored in a vector of length \(2^d-1\).

A shock arrival intensity of zero corresponds to an almost surely infinite shock arrival time.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\), corresponding to the integer \(j\), if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

Simulation algorithms

Exogenous shock model

The exogenous shock model (ESM) simulates a Marshall–Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets of components and defines each component as the minimum of all shock times corresponding to a subset containing this component, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

Arnold model

The Arnold model (AM) simulates a Marshall–Olkin distributed random vector by simulating a marked homogeneous Poisson process where the inter-arrival times correspond to shock inter-arrival times and the marks to the specific shocks; the process is stopped when all components are hit by a shock. As in the ESM, each component is defined as the minimum of all shock times corresponding to a subset containing this component, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07825561 4.43765700 0.07825561
#>  [2,] 0.09380781 0.09380781 0.09380781
#>  [3,] 0.12260547 0.12260547 0.12260547
#>  [4,] 1.31949590 2.81994129 2.81994129
#>  [5,] 0.59522485 0.50485773 0.50485773
#>  [6,] 0.36400962 0.36400962 0.27871695
#>  [7,] 0.36794467 1.08537091 1.08537091
#>  [8,] 0.27754175 1.26309429 0.32702612
#>  [9,] 1.55258767 2.18791035 1.52251721
#> [10,] 0.36677685 0.36677685 0.34944085
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4030790 0.93135608 0.63901786
#>  [2,] 2.7782973 0.14557171 0.61465077
#>  [3,] 1.4331782 0.12344517 0.38621413
#>  [4,] 1.1779799 0.13522784 0.18490875
#>  [5,] 2.1459770 0.94154656 0.37513224
#>  [6,] 1.8587470 0.17585055 0.57339939
#>  [7,] 0.2552004 2.33364044 0.07970708
#>  [8,] 1.7318357 0.02101313 1.95674447
#>  [9,] 0.3930309 6.02251377 1.62409627
#> [10,] 0.3111023 0.40988068 1.93295163
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 5.53315507 5.53315507 5.53315507
#>  [2,] 0.38182246 0.38182246 0.38182246
#>  [3,] 0.09676799 0.09676799 0.09676799
#>  [4,] 1.48366892 1.48366892 1.48366892
#>  [5,] 0.42435098 0.42435098 0.42435098
#>  [6,] 1.58731345 1.58731345 1.58731345
#>  [7,] 0.10384827 0.10384827 0.10384827
#>  [8,] 0.01913130 0.01913130 0.01913130
#>  [9,] 0.85622830 0.85622830 0.85622830
#> [10,] 2.58166418 2.58166418 2.58166418

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.10743497 1.0780457 1.07804570
#>  [2,] 1.39734973 1.3973497 0.14057496
#>  [3,] 0.01104197 3.7135660 1.73526721
#>  [4,] 2.89725366 0.2893662 2.89725366
#>  [5,] 2.87703935 2.8770394 2.87703935
#>  [6,] 1.08243918 0.3791776 1.56272663
#>  [7,] 0.44204078 0.8916500 0.01937754
#>  [8,] 1.04012481 1.3184172 0.91542545
#>  [9,] 0.27712240 0.3166916 0.76166201
#> [10,] 0.39273295 0.3927329 0.39273295
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.46896797 1.4848279 1.5469031
#>  [2,] 0.32798939 0.9222976 0.5338133
#>  [3,] 0.09796986 0.2738552 2.6670164
#>  [4,] 0.50610044 0.5244193 0.4858351
#>  [5,] 0.02126969 0.4591446 0.2799444
#>  [6,] 0.33698835 0.2655862 0.1438801
#>  [7,] 1.75068547 0.7603167 0.9502679
#>  [8,] 1.01621156 1.9688929 0.5530013
#>  [9,] 1.81956823 0.1230909 1.7002348
#> [10,] 0.53508500 0.1660258 0.5888802
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.08325089 2.08325089 2.08325089
#>  [2,] 1.48869664 1.48869664 1.48869664
#>  [3,] 0.74384496 0.74384496 0.74384496
#>  [4,] 0.28881347 0.28881347 0.28881347
#>  [5,] 1.50477493 1.50477493 1.50477493
#>  [6,] 0.31159245 0.31159245 0.31159245
#>  [7,] 0.40380966 0.40380966 0.40380966
#>  [8,] 0.54078589 0.54078589 0.54078589
#>  [9,] 0.03597492 0.03597492 0.03597492
#> [10,] 1.23657122 1.23657122 1.23657122

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2345755 0.23457550 0.28099858
#>  [2,] 1.1608626 0.98491808 1.16086259
#>  [3,] 0.5022832 0.50228319 0.50228319
#>  [4,] 0.5660747 0.56607470 0.04163709
#>  [5,] 2.3834030 1.83996331 0.56117668
#>  [6,] 1.2803742 1.78399338 1.78399338
#>  [7,] 0.1814155 0.23553593 0.23553593
#>  [8,] 1.2353518 0.38790944 2.19926622
#>  [9,] 0.0215896 0.06879133 0.06879133
#> [10,] 0.1220778 2.74465426 0.12207779
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.7612576 4.13816564 1.58027365
#>  [2,] 0.2051282 0.29246895 0.08712489
#>  [3,] 1.3816173 0.07742762 3.07586641
#>  [4,] 0.6297019 1.60640893 0.51295926
#>  [5,] 1.0456900 1.85510149 0.50539882
#>  [6,] 2.8503032 1.31493504 0.41391256
#>  [7,] 0.2302979 0.27753585 3.69762410
#>  [8,] 0.6554698 1.48939886 0.35125505
#>  [9,] 0.3424783 0.21972510 2.51311670
#> [10,] 0.1504838 1.50073851 1.07825594
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.81080848 2.81080848 2.81080848
#>  [2,] 0.39776922 0.39776922 0.39776922
#>  [3,] 0.06217128 0.06217128 0.06217128
#>  [4,] 0.49916999 0.49916999 0.49916999
#>  [5,] 1.02573071 1.02573071 1.02573071
#>  [6,] 4.67057120 4.67057120 4.67057120
#>  [7,] 0.80261334 0.80261334 0.80261334
#>  [8,] 0.10170462 0.10170462 0.10170462
#>  [9,] 0.04364553 0.04364553 0.04364553
#> [10,] 0.01735275 0.01735275 0.01735275