Sample from a Marshall--Olkin distribution

Draws n independent samples from a d-variate Marshall-Olkin distribution with shock rates intensities.

rmo_esm(n, d, intensities)

rmo_arnold(n, d, intensities)

Arguments

n
Number of samples

d
Dimension

intensities
Marshall-Olkin intensity rates

Value

rmo_esm implements the exogenous shock model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

rmo_arnold implements the Arnold model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

Details

The shock intensities:

The shock intensities must be stored in a vector of length \(2^d-1\).

A shock intensity of zero corresponds to an almost surely infinite shock.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\) corresponding to the integer \(j\) if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

The exogenous shock model simulates a Marshall--Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

The Arnold model simulates a Marshall--Olkin distributed random variable by simulating a marked homogeneous Poisson process and where the inter-arrival times correspond to shock shock-arrival times and the marks to the specific shocks, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 , https://doi.org/10.1142/10265.

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 , https://doi.org/10.2307/2282907.

See also

Examples

rmo_esm(10L, 2L, c(0.4, 0.3, 0.2))

#>             [,1]      [,2]
#>  [1,] 0.26680386 0.2668039
#>  [2,] 5.62700193 1.2437947
#>  [3,] 0.55773144 0.2325658
#>  [4,] 0.05561243 0.4859210
#>  [5,] 0.42442786 1.0249917
#>  [6,] 0.44918778 1.7277225
#>  [7,] 1.31409735 4.5407563
#>  [8,] 1.02003859 1.0200386
#>  [9,] 0.78497915 2.4786463
#> [10,] 1.16914558 1.1691456

rmo_esm(10L, 2L, c(1, 1, 0))         ## independence

#>             [,1]       [,2]
#>  [1,] 0.19694999 1.32056170
#>  [2,] 0.35882925 3.55473991
#>  [3,] 0.34689565 1.17822671
#>  [4,] 0.46104855 0.51208218
#>  [5,] 0.08567431 1.67551998
#>  [6,] 0.45389771 0.27537110
#>  [7,] 1.27878700 0.03479680
#>  [8,] 0.15237484 1.47353128
#>  [9,] 0.60255025 0.51610724
#> [10,] 0.06513033 0.09360955

rmo_esm(10L, 2L, c(0, 0, 1))         ## comonotone

#>            [,1]      [,2]
#>  [1,] 2.6142656 2.6142656
#>  [2,] 1.2465462 1.2465462
#>  [3,] 1.7651095 1.7651095
#>  [4,] 0.3814517 0.3814517
#>  [5,] 0.3504170 0.3504170
#>  [6,] 0.1360090 0.1360090
#>  [7,] 1.2415323 1.2415323
#>  [8,] 0.3006878 0.3006878
#>  [9,] 0.6291924 0.6291924
#> [10,] 2.2148779 2.2148779

rmo_arnold(10L, 2L, c(0.4, 0.3, 0.2))

#>             [,1]       [,2]
#>  [1,] 3.79427775 0.05962294
#>  [2,] 5.40393171 2.87103046
#>  [3,] 3.16795179 0.72778092
#>  [4,] 0.03412289 0.03412289
#>  [5,] 1.94384269 3.85398022
#>  [6,] 1.97734914 0.63544703
#>  [7,] 0.23055645 0.23055645
#>  [8,] 0.51757926 4.80876075
#>  [9,] 1.38119456 1.73471327
#> [10,] 1.07777561 0.45784087

rmo_arnold(10L, 2L, c(1, 1, 0))         ## independence

#>            [,1]         [,2]
#>  [1,] 0.1101377 8.207787e-01
#>  [2,] 1.0515201 6.963041e-01
#>  [3,] 0.4092766 3.593669e-05
#>  [4,] 1.0909632 1.392690e-03
#>  [5,] 1.0104634 1.627403e+00
#>  [6,] 1.3111097 5.739979e-01
#>  [7,] 0.3731909 4.131125e-01
#>  [8,] 0.3819071 1.776787e+00
#>  [9,] 0.4645737 2.038125e-01
#> [10,] 2.3415224 1.044819e+00

rmo_arnold(10L, 2L, c(0, 0, 1))         ## comonotone

#>             [,1]       [,2]
#>  [1,] 1.30110875 1.30110875
#>  [2,] 0.05664471 0.05664471
#>  [3,] 1.97830060 1.97830060
#>  [4,] 0.50047969 0.50047969
#>  [5,] 0.03178032 0.03178032
#>  [6,] 0.13373812 0.13373812
#>  [7,] 1.25931561 1.25931561
#>  [8,] 0.34743799 0.34743799
#>  [9,] 1.15793598 1.15793598
#> [10,] 0.10258134 0.10258134