Sample from a Marshall--Olkin distribution

Draws n independent samples from a d-variate Marshall-Olkin distribution with shock rates intensities.

rmo_esm(n, d, intensities)

rmo_arnold(n, d, intensities)

Arguments

n
Number of samples

d
Dimension

intensities
Marshall-Olkin shock intensities

Value

rmo_esm implements the exogenous shock model algorithm and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

rmo_arnold implements the Arnold model algorithm and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

Details

Parameterisation: The Marshall--Olkin distribution has the survival function $$ \bar{F}(t) \bar{F}(t) = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 , $$ for shock intensities \(\lambda_I \geq 0\), \(\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}\).

The shock intensities:

The shock intensities must be stored in a vector of length \(2^d-1\).

A shock intensity of zero corresponds to an almost surely infinite shock.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\) corresponding to the integer \(j\) if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

The exogenous shock model simulates a Marshall--Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

The Arnold model simulates a Marshall--Olkin distributed random variable by simulating a marked homogeneous Poisson process and where the inter-arrival times correspond to shock shock-arrival times and the marks to the specific shocks, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 , https://doi.org/10.1142/10265.

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 , https://doi.org/10.2307/2282907.

See also

Examples

rmo_esm(10L, 2L, c(0.4, 0.3, 0.2))

#>             [,1]       [,2]
#>  [1,] 0.02485454 1.22460525
#>  [2,] 0.04444998 0.04444998
#>  [3,] 2.93012918 0.66208227
#>  [4,] 1.60133471 4.72094217
#>  [5,] 0.45652077 0.45652077
#>  [6,] 1.52258871 1.45663420
#>  [7,] 0.85053079 1.93314782
#>  [8,] 3.03242032 3.03242032
#>  [9,] 0.35837738 0.35837738
#> [10,] 0.01374046 0.01374046

rmo_esm(10L, 2L, c(1, 1, 0))         ## independence

#>            [,1]       [,2]
#>  [1,] 0.3063407 0.08930288
#>  [2,] 1.6969126 0.56514642
#>  [3,] 0.3622123 0.28274297
#>  [4,] 0.2622528 0.08417149
#>  [5,] 2.9578325 3.32939341
#>  [6,] 0.5888173 1.92909570
#>  [7,] 0.1175686 0.52503279
#>  [8,] 0.2629203 0.46106682
#>  [9,] 0.7315058 0.23970906
#> [10,] 1.5693407 1.80407874

rmo_esm(10L, 2L, c(0, 0, 1))         ## comonotone

#>            [,1]      [,2]
#>  [1,] 2.4483525 2.4483525
#>  [2,] 0.2120679 0.2120679
#>  [3,] 0.1135205 0.1135205
#>  [4,] 0.2083842 0.2083842
#>  [5,] 1.3475713 1.3475713
#>  [6,] 0.3741313 0.3741313
#>  [7,] 2.1784409 2.1784409
#>  [8,] 2.6780586 2.6780586
#>  [9,] 1.4470229 1.4470229
#> [10,] 1.2944415 1.2944415

rmo_arnold(10L, 2L, c(0.4, 0.3, 0.2))

#>            [,1]      [,2]
#>  [1,] 2.4185023 0.9627006
#>  [2,] 1.8170382 2.3417722
#>  [3,] 3.6807703 3.6807703
#>  [4,] 0.3787061 3.7632844
#>  [5,] 0.1393665 0.1393665
#>  [6,] 0.5207416 0.5207416
#>  [7,] 0.2767178 0.2767178
#>  [8,] 0.8146618 5.3949811
#>  [9,] 0.6663180 1.0934350
#> [10,] 0.2835903 1.6765077

rmo_arnold(10L, 2L, c(1, 1, 0))         ## independence

#>              [,1]      [,2]
#>  [1,] 0.004444643 0.8787443
#>  [2,] 0.353407168 1.6469284
#>  [3,] 3.960596476 2.6962909
#>  [4,] 0.394150515 0.4163511
#>  [5,] 1.160507596 0.3122503
#>  [6,] 0.981750977 2.3252934
#>  [7,] 0.537513839 0.5738957
#>  [8,] 2.107442802 0.9459082
#>  [9,] 0.114444256 2.0515040
#> [10,] 0.838840026 0.5401260

rmo_arnold(10L, 2L, c(0, 0, 1))         ## comonotone

#>            [,1]      [,2]
#>  [1,] 1.5415030 1.5415030
#>  [2,] 0.3498185 0.3498185
#>  [3,] 2.0447361 2.0447361
#>  [4,] 0.2159389 0.2159389
#>  [5,] 0.9881716 0.9881716
#>  [6,] 0.5351129 0.5351129
#>  [7,] 2.3410853 2.3410853
#>  [8,] 1.0399577 1.0399577
#>  [9,] 1.1035778 1.1035778
#> [10,] 0.3760116 0.3760116