Sample from a Marshall--Olkin distribution

Draws n independent samples from a d-variate Marshall-Olkin distribution with shock rates intensities.

rmo_esm(n, d, intensities)

rmo_arnold(n, d, intensities)

Arguments

n
Number of samples

d
Dimension

intensities
Marshall-Olkin intensity rates

Value

rmo_esm implements the exogenous shock model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

rmo_arnold implements the Arnold model representation and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of a \(d\) variate Marshall-Olkin distribution with parameters intensities.

Details

The shock intensities:

The shock intensities must be stored in a vector of length \(2^d-1\).

A shock intensity of zero corresponds to an almost surely infinite shock.

We use a binary representation to map a non-empty subset \(J\) of \(\{ 1, \ldots, d\}\) to integers \(j\) of \(1, \ldots, 2^d-1\). In particular, \(i\) is a component in the set \(J\) corresponding to the integer \(j\) if, and only if, \(j = \sum_{k=0}^\infty a_k 2^k\) and \(a_{i-1} = 1\).

The exogenous shock model simulates a Marshall--Olkin distributed random vector via exponentially distributed shock times for all non-empty subsets, see (see pp. 104 psqq. Mai and Scherer 2017) and (Marshall and Olkin 1967) .

The Arnold model simulates a Marshall--Olkin distributed random variable by simulating a marked homogeneous Poisson process and where the inter-arrival times correspond to shock shock-arrival times and the marks to the specific shocks, see (see Sec. 3.1.2 Mai and Scherer 2017) and (Arnold 1975) .

References

Arnold BC (1975). “A characterization of the exponential distribution by multivariate geometric compounding.” Sankhy\=a: The Indian Journal of Statistics, Series A, 37(1), 164--173.

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

Marshall AW, Olkin I (1967). “A multivariate exponential distribution.” Journal of the American Statistical Association, 62(317), 30--44. doi: 10.2307/2282907 .

See also

Examples

rmo_esm(10L, 2L, c(0.4, 0.3, 0.2))

#>            [,1]      [,2]
#>  [1,] 1.0076089 1.0076089
#>  [2,] 4.1099047 4.2610808
#>  [3,] 0.1186714 0.1186714
#>  [4,] 2.4367253 4.3415338
#>  [5,] 3.2682491 6.2397622
#>  [6,] 0.1782992 0.1782992
#>  [7,] 0.9791194 1.2570400
#>  [8,] 1.0321692 0.4166710
#>  [9,] 1.3488047 1.3488047
#> [10,] 0.3157354 0.3157354

rmo_esm(10L, 2L, c(1, 1, 0))         ## independence

#>            [,1]      [,2]
#>  [1,] 0.8346728 2.1228993
#>  [2,] 1.1099744 0.6503988
#>  [3,] 0.7884202 0.1400899
#>  [4,] 1.0360235 0.1925256
#>  [5,] 1.9184386 0.3760188
#>  [6,] 0.8075358 0.7125976
#>  [7,] 0.7070553 1.2010309
#>  [8,] 0.1198257 0.2667298
#>  [9,] 0.2012750 0.7386326
#> [10,] 0.8533478 0.9709052

rmo_esm(10L, 2L, c(0, 0, 1))         ## comonotone

#>             [,1]       [,2]
#>  [1,] 1.42692257 1.42692257
#>  [2,] 0.10330351 0.10330351
#>  [3,] 1.76547593 1.76547593
#>  [4,] 3.79108831 3.79108831
#>  [5,] 0.62810262 0.62810262
#>  [6,] 1.30879119 1.30879119
#>  [7,] 1.67628039 1.67628039
#>  [8,] 0.01530056 0.01530056
#>  [9,] 0.32071861 0.32071861
#> [10,] 0.02358263 0.02358263

rmo_arnold(10L, 2L, c(0.4, 0.3, 0.2))

#>               [,1]        [,2]
#>  [1,] 0.7545132037 0.745672082
#>  [2,] 0.0004255575 0.611343385
#>  [3,] 0.2977252690 0.297725269
#>  [4,] 0.1486540860 3.461766826
#>  [5,] 2.0001106566 0.425421080
#>  [6,] 0.2485609277 3.826419996
#>  [7,] 0.4477469270 0.447746927
#>  [8,] 0.4688249994 1.424578000
#>  [9,] 0.0034628925 0.003462892
#> [10,] 1.2844908831 2.504649150

rmo_arnold(10L, 2L, c(1, 1, 0))         ## independence

#>            [,1]      [,2]
#>  [1,] 0.5942469 0.1642855
#>  [2,] 0.5251676 0.4902827
#>  [3,] 0.2838528 0.2727304
#>  [4,] 0.7068440 1.4998866
#>  [5,] 0.2044221 0.8855356
#>  [6,] 1.1864805 2.0846265
#>  [7,] 0.1169146 4.5708246
#>  [8,] 0.0173984 1.0590224
#>  [9,] 0.8139989 0.6232731
#> [10,] 0.2183484 0.0680045

rmo_arnold(10L, 2L, c(0, 0, 1))         ## comonotone

#>             [,1]       [,2]
#>  [1,] 2.21487791 2.21487791
#>  [2,] 0.52694965 0.52694965
#>  [3,] 0.10494467 0.10494467
#>  [4,] 3.58954107 3.58954107
#>  [5,] 0.67289250 0.67289250
#>  [6,] 0.02489078 0.02489078
#>  [7,] 0.39316732 0.39316732
#>  [8,] 2.12962777 2.12962777
#>  [9,] 0.14998335 0.14998335
#> [10,] 0.09718288 0.09718288