Sample from Cuadras-Auge distribution

Draws n independent samples from a d variate Cuadras-Augé distribution with parameters alpha and beta.

rcamo_esm(n, d, alpha, beta)

Arguments

n
Number of samples

d
Dimension

alpha
Shock intensity of individual shocks

beta
Shock intensity of global shock

Value

rmo_esm_cuadras_auge implements an optimized version of the exogenous shock model algorithm for the Cuadras-Augé family and returns an \(n \times d\) array matrix with rows corresponding to the independent samples of size \(d\).

Details

Parameterisation: The Cuadras-Augé distribution is a special case of the exchangeable Marshall-Olkin distribution where \(\lambda_i = 0 \forall 1 < i < d\). The individual shock rate is \(\alpha = \lambda_1\) and the global shock rate is \(\beta = \lambda_d\).

See also

Examples

rcamo_esm(10L, 2L, 0.5, 0.2)

#>            [,1]      [,2]
#>  [1,] 1.1591114 0.5491868
#>  [2,] 4.3117345 2.5761153
#>  [3,] 1.4992099 0.7403792
#>  [4,] 2.2715207 0.9507816
#>  [5,] 0.1075008 0.1075008
#>  [6,] 0.3422821 1.1239411
#>  [7,] 1.2966656 3.7413400
#>  [8,] 0.2143431 0.2143431
#>  [9,] 5.6632929 1.6167282
#> [10,] 0.7738252 1.2734566

rcamo_esm(10L, 2L, 0, 1)      ## comonotone

#>             [,1]       [,2]
#>  [1,] 0.01923201 0.01923201
#>  [2,] 0.30315226 0.30315226
#>  [3,] 0.23228084 0.23228084
#>  [4,] 0.19145614 0.19145614
#>  [5,] 0.40272523 0.40272523
#>  [6,] 0.97748519 0.97748519
#>  [7,] 0.14814172 0.14814172
#>  [8,] 4.35295840 4.35295840
#>  [9,] 0.52214073 0.52214073
#> [10,] 0.37999728 0.37999728

rcamo_esm(10L, 2L, 1, 0)      ## independence

#>            [,1]        [,2]
#>  [1,] 0.2929692 1.602440214
#>  [2,] 0.3148262 0.599230511
#>  [3,] 0.2593448 0.479714277
#>  [4,] 0.2089334 0.864223083
#>  [5,] 2.3641674 1.118781199
#>  [6,] 1.1572743 0.350783662
#>  [7,] 3.0920279 1.256454813
#>  [8,] 0.1299232 0.313174729
#>  [9,] 0.2345208 0.006493014
#> [10,] 0.1396040 1.135673976