Sample from an exchangeable MO distribution

Draws n independent samples from a d variate exchangeable Marshall-Olkin distribution with shock rates ex_intensities.

rmo_ex_arnold(n, d, ex_intensities)

Arguments

n
number of samples

d
dimension

ex_intensities
exchangeable Marshall-Olkin intensity rates

Value

rmo_ex_arnold implements the modified Arnold model for the exchangeable subclass and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically disctributed samples of a \(d\) variate exchangeable Marshall-Olkin distribution with exchangeable parameters ex_intensities.

Details

The exchangeable shock intensities must be stored in a vector of length \(d\).

The entry \({exintensities}_{i}\) is the intensity of a shock corresponding to a set with \(i\) elements.

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 122 psqq.

See also

Examples

rmo_ex_arnold(10, 2, c(0.4, 0.2))

#>             [,1]      [,2]
#>  [1,] 0.68832679 2.6866876
#>  [2,] 2.47438250 2.4743825
#>  [3,] 3.60663403 2.3336104
#>  [4,] 0.42350683 0.4235068
#>  [5,] 2.25078765 2.2507877
#>  [6,] 2.43157692 1.1624633
#>  [7,] 1.65857297 2.3379480
#>  [8,] 4.00425389 0.5215223
#>  [9,] 2.59340590 3.5016621
#> [10,] 0.05664471 1.2868070

rmo_ex_arnold(10, 2, c(1, 0))      ## independence

#>            [,1]       [,2]
#>  [1,] 0.2502398 0.28202017
#>  [2,] 0.1196705 0.06686906
#>  [3,] 1.1591830 0.00124707
#>  [4,] 1.0059825 1.92822463
#>  [5,] 2.7549718 0.73936227
#>  [6,] 1.2614949 0.47776296
#>  [7,] 0.0594912 2.35306153
#>  [8,] 0.6784965 2.52381421
#>  [9,] 0.2657598 0.02016214
#> [10,] 0.4410851 0.35167474

rmo_ex_arnold(10, 2, c(0, 1))      ## comonotone

#>            [,1]      [,2]
#>  [1,] 0.3716193 0.3716193
#>  [2,] 0.2781220 0.2781220
#>  [3,] 0.8172443 0.8172443
#>  [4,] 0.6558426 0.6558426
#>  [5,] 0.8203975 0.8203975
#>  [6,] 0.4683331 0.4683331
#>  [7,] 0.4498702 0.4498702
#>  [8,] 3.1486301 3.1486301
#>  [9,] 0.6861939 0.6861939
#> [10,] 1.4543557 1.4543557