Sample with LFM and CPP subordinator

Draws n independent samples from a d-variate extendible Marshall-Olkin distribution corresponding to a LFM with a compound Poisson subordinator with parameters rate, rate_killing, rate_drift, rjump_name, and rjump_arg_list.

rmo_lfm_cpp(
  n,
  d,
  rate,
  rate_killing,
  rate_drift,
  rjump_name,
  rjump_arg_list = list()
)

Arguments

n
number of samples

d
dimension

rate
rate of CPP subordinator

rate_killing
killing rate of CPP subordinator

rate_drift
drift rate of CPP subordinator

rjump_name
name of jump sampling function for jumps of CPP subordinator

rjump_arg_list
list with named arguments for jump sampling function for jumps of CPP subordinator

Value

rmo_lfm_cpp implements the Lévy-frailty model representation with a compound Poisson subordinator and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of the corresponding d-variate extendible Marshall-Olkin distribution.

Details

rate is the jump intensity of the compound Poisson subordinator: at each given point-in-time, the waiting time to the next jump is exponentially distributed with rate rate.

rate_killing is the killing intensity of the compound Poisson subordinator: the probability that the compound Poisson subordinator jumps to its graveyard-state \(\infty\) between \(t\) and \(s\) is \(1 - \exp{\{-{ratekilling} (s-t)\}}\).

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 140 psqq.

See also

Examples

rmo_lfm_cpp(10L, 2L, 0.5, 0.1, 0.2, "rposval", list("value"=1))

#>            [,1]      [,2]
#>  [1,] 3.0695363 1.2506929
#>  [2,] 2.1417266 2.9253122
#>  [3,] 0.1943344 0.4411985
#>  [4,] 1.4795995 0.6118691
#>  [5,] 0.5381771 0.4922071
#>  [6,] 1.8323667 4.3971396
#>  [7,] 0.2061563 0.2061563
#>  [8,] 2.4362132 2.7844158
#>  [9,] 0.1776062 0.1776062
#> [10,] 0.2341490 0.1096488

rmo_lfm_cpp(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))

#>             [,1]       [,2]
#>  [1,]  1.6269882  4.6085298
#>  [2,]  6.1571251  6.4251133
#>  [3,]  8.0175440  2.3281082
#>  [4,] 10.4896824 13.3182855
#>  [5,] 17.2426682 18.9168472
#>  [6,]  3.3909483  1.2996616
#>  [7,]  4.1571849 11.1076425
#>  [8,]  1.3897280  1.3897280
#>  [9,]  0.4064594  1.2095294
#> [10,]  1.4631810  0.9971864

rmo_lfm_cpp(10L, 2L, 0, 0, 1, "rposval", list("value"=1))  ## independence

#>             [,1]       [,2]
#>  [1,] 1.29953920 6.92432903
#>  [2,] 1.36501132 0.52786872
#>  [3,] 0.58895533 1.40646782
#>  [4,] 0.06081897 1.25410909
#>  [5,] 0.35004695 2.58293262
#>  [6,] 1.96600090 1.06611532
#>  [7,] 0.15579552 0.38183566
#>  [8,] 0.94246480 0.77757018
#>  [9,] 0.95607280 0.63237537
#> [10,] 0.09183088 0.02670447

rmo_lfm_cpp(10L, 2L, 0, 1, 0, "rposval", list("value"=1))  ## comonotone

#>             [,1]       [,2]
#>  [1,] 2.13003743 2.13003743
#>  [2,] 0.38143820 0.38143820
#>  [3,] 2.66359687 2.66359687
#>  [4,] 1.27157350 1.27157350
#>  [5,] 0.07748686 0.07748686
#>  [6,] 0.70599219 0.70599219
#>  [7,] 0.30400037 0.30400037
#>  [8,] 0.69140711 0.69140711
#>  [9,] 1.01687513 1.01687513
#> [10,] 1.25896015 1.25896015