Sample with LFM and CPP subordinator

Draws n independent samples from a d-variate extendible Marshall-Olkin distribution corresponding to a LFM with a compound Poisson subordinator with parameters rate, rate_killing, rate_drift, rjump_name, and rjump_arg_list.

rmo_lfm_cpp(
  n,
  d,
  rate,
  rate_killing,
  rate_drift,
  rjump_name,
  rjump_arg_list = list()
)

Arguments

n
number of samples

d
dimension

rate
rate of CPP subordinator

rate_killing
killing rate of CPP subordinator

rate_drift
drift rate of CPP subordinator

rjump_name
name of jump sampling function for jumps of CPP subordinator

rjump_arg_list
list with named arguments for jump sampling function for jumps of CPP subordinator

Value

rmo_lfm_cpp implements the Lévy-frailty model representation with a compound Poisson subordinator and returns an \(n \times d\) numeric matrix with the rows corresponding to independent and identically distributed samples of the corresponding d-variate extendible Marshall-Olkin distribution.

Details

rate is the jump intensity of the compound Poisson subordinator: at each given point-in-time, the waiting time to the next jump is exponentially distributed with rate rate.

rate_killing is the killing intensity of the compound Poisson subordinator: the probability that the compound Poisson subordinator jumps to its graveyard-state \(\infty\) between \(t\) and \(s\) is \(1 - \exp{\{-{ratekilling} (s-t)\}}\).

References

For more information on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 140 psqq.

See also

Examples

rmo_lfm_cpp(10L, 2L, 0.5, 0.1, 0.2, "rposval", list("value"=1))

#>            [,1]      [,2]
#>  [1,] 0.7231945 0.7231945
#>  [2,] 1.0434329 1.9678475
#>  [3,] 0.3917928 0.1943344
#>  [4,] 2.7146061 3.2672105
#>  [5,] 0.9693047 0.7878149
#>  [6,] 0.4922071 0.7329467
#>  [7,] 2.0932052 2.0932052
#>  [8,] 0.5153908 4.8724263
#>  [9,] 0.1776062 0.1776062
#> [10,] 0.1751708 0.2192977

rmo_lfm_cpp(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))

#>              [,1]        [,2]
#>  [1,]  4.77753984  4.77753984
#>  [2,]  9.11884536  5.15626252
#>  [3,]  0.48587751  0.48587751
#>  [4,]  2.49043851  2.31018646
#>  [5,]  1.34824690  1.34824690
#>  [6,] 11.28755170  4.01577216
#>  [7,]  2.21402669  1.61283188
#>  [8,]  0.01210665  5.14202710
#>  [9,]  9.28832798  0.01358033
#> [10,] 20.81885859 38.53618972

rmo_lfm_cpp(10L, 2L, 0, 0, 1, "rposval", list("value"=1))  ## independence

#>             [,1]       [,2]
#>  [1,] 1.25410909 0.35004695
#>  [2,] 2.58293262 1.96600090
#>  [3,] 1.06611532 0.15579552
#>  [4,] 0.38183566 0.94246480
#>  [5,] 0.77757018 0.95607280
#>  [6,] 0.63237537 0.09183088
#>  [7,] 0.02670447 2.35334572
#>  [8,] 0.89676529 2.13003743
#>  [9,] 0.14312169 0.35597612
#> [10,] 0.38143820 1.18679047

rmo_lfm_cpp(10L, 2L, 0, 1, 0, "rposval", list("value"=1))  ## comonotone

#>             [,1]       [,2]
#>  [1,] 0.54960317 0.54960317
#>  [2,] 0.44061157 0.44061157
#>  [3,] 0.03920342 0.03920342
#>  [4,] 0.11517457 0.11517457
#>  [5,] 0.28590440 0.28590440
#>  [6,] 0.02643718 0.02643718
#>  [7,] 1.80836899 1.80836899
#>  [8,] 0.09659652 0.09659652
#>  [9,] 0.47850159 0.47850159
#> [10,] 0.34771273 0.34771273