Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.38471197 0.20731834 2.9602781
#>  [2,] 1.67353413 0.74851294 1.9844646
#>  [3,] 0.69963092 0.31400621 0.2887767
#>  [4,] 1.05041472 0.52848169 1.1056045
#>  [5,] 0.42742110 0.46145842 0.3780742
#>  [6,] 0.80495712 0.07111229 1.7097449
#>  [7,] 0.40407743 0.40407743 0.4040774
#>  [8,] 0.43703500 4.54929801 0.6853945
#>  [9,] 0.05736984 0.33865201 1.7524649
#> [10,] 5.88828941 0.47482872 0.3125398
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7440582 0.7440582 0.7440582
#>  [2,] 0.6000785 0.6000785 0.6000785
#>  [3,] 0.7660618 0.7660618 0.7660618
#>  [4,] 1.3351624 1.3351624 1.3351624
#>  [5,] 0.5348441 0.5348441 0.5348441
#>  [6,] 0.3075862 0.3075862 0.3075862
#>  [7,] 0.3325900 0.3325900 0.3325900
#>  [8,] 0.9953178 0.9953178 0.9953178
#>  [9,] 0.8523534 0.8523534 0.8523534
#> [10,] 0.2409691 0.2409691 0.2409691
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]      [,3]
#>  [1,] 2.22952091 0.62293594 0.5666936
#>  [2,] 0.39689975 1.68804320 1.4201404
#>  [3,] 0.01957947 0.07091638 0.7394691
#>  [4,] 0.39679123 0.41680343 0.5872069
#>  [5,] 0.07305307 0.45687146 0.2157368
#>  [6,] 0.84192264 0.05475117 0.1594599
#>  [7,] 0.37397843 0.07973184 0.9618996
#>  [8,] 0.80990773 1.34804995 3.7896139
#>  [9,] 0.53529786 0.75465080 0.1674957
#> [10,] 0.74658698 1.46148864 1.8241370

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.12967719 0.9136108 3.6899728
#>  [2,] 0.19667624 1.8933146 0.4997865
#>  [3,] 0.48842146 1.2317957 0.4221915
#>  [4,] 0.31226863 0.3122686 0.3122686
#>  [5,] 1.09060468 1.0906047 1.0906047
#>  [6,] 0.01959357 4.2545130 9.1017504
#>  [7,] 0.33795951 6.9331961 0.7630142
#>  [8,] 0.75724063 0.2365362 0.2231161
#>  [9,] 0.09514877 2.7447088 2.1297206
#> [10,] 0.15758499 0.1575850 0.1575850
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.50443018 1.50443018 1.50443018
#>  [2,] 0.37022875 0.37022875 0.37022875
#>  [3,] 0.00550371 0.00550371 0.00550371
#>  [4,] 0.21291412 0.21291412 0.21291412
#>  [5,] 0.05302529 0.05302529 0.05302529
#>  [6,] 1.21599380 1.21599380 1.21599380
#>  [7,] 0.18490333 0.18490333 0.18490333
#>  [8,] 1.28994512 1.28994512 1.28994512
#>  [9,] 1.41940260 1.41940260 1.41940260
#> [10,] 0.06308340 0.06308340 0.06308340
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3334477 0.20396960 0.82968189
#>  [2,] 0.6040270 0.63544106 0.98403135
#>  [3,] 0.6847461 0.68824534 0.76364963
#>  [4,] 0.4897941 0.75430847 4.94541121
#>  [5,] 0.2684092 0.45170277 0.11942586
#>  [6,] 1.0722526 0.88415652 0.62669852
#>  [7,] 3.4892261 0.05394067 3.17129192
#>  [8,] 1.2590574 0.38286813 0.55513766
#>  [9,] 0.3590597 0.08064619 0.09747055
#> [10,] 0.5413360 1.35205028 1.04728624

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.40073260 0.8394852 3.4960273
#>  [2,] 1.50258274 1.1243828 1.5025827
#>  [3,] 0.12317176 0.3109686 0.2616707
#>  [4,] 0.02323719 0.6902630 5.1878879
#>  [5,] 3.86808898 1.2383904 6.0715860
#>  [6,] 1.47506659 1.6068274 1.5302060
#>  [7,] 0.40352425 0.4035242 0.4035242
#>  [8,] 1.49227410 0.7186343 1.8602204
#>  [9,] 3.37919270 5.1676252 0.5879220
#> [10,] 2.84378461 2.9785374 1.3707562
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.1900679 3.1900679 3.1900679
#>  [2,] 0.5493546 0.5493546 0.5493546
#>  [3,] 0.5419027 0.5419027 0.5419027
#>  [4,] 1.2203506 1.2203506 1.2203506
#>  [5,] 1.0736953 1.0736953 1.0736953
#>  [6,] 1.1288912 1.1288912 1.1288912
#>  [7,] 0.3286512 0.3286512 0.3286512
#>  [8,] 0.2262894 0.2262894 0.2262894
#>  [9,] 2.0906233 2.0906233 2.0906233
#> [10,] 1.6766829 1.6766829 1.6766829
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>              [,1]      [,2]       [,3]
#>  [1,] 2.453890623 0.1909259 1.15969605
#>  [2,] 0.301680794 1.0913022 1.40913584
#>  [3,] 2.597364502 0.2947611 0.91085329
#>  [4,] 1.172863951 0.1569631 1.31818445
#>  [5,] 1.668380365 0.3601119 0.56759589
#>  [6,] 1.177127064 3.3326799 0.09515901
#>  [7,] 0.008021454 2.2744881 1.83868175
#>  [8,] 0.140315005 0.8112638 0.45710222
#>  [9,] 1.860641558 0.5340296 1.05314488
#> [10,] 0.666254909 1.2429205 0.66984586

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0033744 0.3290450 1.18590950
#>  [2,] 1.6670095 0.4923779 1.31402315
#>  [3,] 3.5068882 0.9659918 1.14706072
#>  [4,] 0.6385405 4.0958668 0.64230797
#>  [5,] 0.5652147 1.1205105 1.18592234
#>  [6,] 0.3257829 0.5570626 1.35209713
#>  [7,] 1.3569552 1.0158805 2.54938416
#>  [8,] 0.3762098 0.6226921 2.29164153
#>  [9,] 0.3365336 0.7892566 0.02516818
#> [10,] 0.3289969 2.1078979 2.04257481
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.01267864 2.01267864 2.01267864
#>  [2,] 0.49944199 0.49944199 0.49944199
#>  [3,] 1.52455806 1.52455806 1.52455806
#>  [4,] 0.98576954 0.98576954 0.98576954
#>  [5,] 0.02487981 0.02487981 0.02487981
#>  [6,] 1.87540697 1.87540697 1.87540697
#>  [7,] 0.14250697 0.14250697 0.14250697
#>  [8,] 2.08257092 2.08257092 2.08257092
#>  [9,] 1.78726457 1.78726457 1.78726457
#> [10,] 1.39869636 1.39869636 1.39869636
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]       [,2]        [,3]
#>  [1,] 1.1596149 2.14558050 2.149955385
#>  [2,] 0.6693484 2.21734387 0.001558413
#>  [3,] 0.3299316 4.52110744 0.863144698
#>  [4,] 0.7606489 0.50313347 1.086675234
#>  [5,] 0.8941757 5.30383198 1.278870221
#>  [6,] 1.0082215 0.32543690 0.616818236
#>  [7,] 0.6213128 0.53664950 3.705817925
#>  [8,] 0.4172451 3.54443384 0.218520753
#>  [9,] 0.6575757 0.06891373 0.611040098
#> [10,] 1.8742000 0.29442278 0.003890466

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.01562861 1.2131822 0.2682724
#>  [2,] 0.24467040 0.2446704 0.2446704
#>  [3,] 4.92790782 1.5932149 2.0986489
#>  [4,] 1.23696686 5.7808650 1.1561981
#>  [5,] 1.50726366 0.4033717 0.1003695
#>  [6,] 1.29582440 1.9860577 2.3948853
#>  [7,] 0.04053248 2.3971474 0.2020230
#>  [8,] 1.41338252 1.5449278 1.5963903
#>  [9,] 5.06283127 0.8252524 0.4694468
#> [10,] 0.15230582 0.1523058 0.1523058
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.238147218 0.238147218 0.238147218
#>  [2,] 0.471528715 0.471528715 0.471528715
#>  [3,] 0.244056945 0.244056945 0.244056945
#>  [4,] 1.054763588 1.054763588 1.054763588
#>  [5,] 2.481132286 2.481132286 2.481132286
#>  [6,] 2.136479935 2.136479935 2.136479935
#>  [7,] 0.006801576 0.006801576 0.006801576
#>  [8,] 0.498688027 0.498688027 0.498688027
#>  [9,] 1.719398549 1.719398549 1.719398549
#> [10,] 1.695359627 1.695359627 1.695359627
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>              [,1]        [,2]      [,3]
#>  [1,] 0.146504722 0.298277867 0.7546487
#>  [2,] 0.642978721 0.191796622 0.8971850
#>  [3,] 0.867651423 0.029251238 3.8233143
#>  [4,] 2.210574043 0.003849827 1.0232574
#>  [5,] 0.830648352 0.481304632 0.7473174
#>  [6,] 0.036847226 3.403523872 1.7518384
#>  [7,] 2.462644953 0.731033066 0.4714331
#>  [8,] 0.004253095 1.500977672 2.2016988
#>  [9,] 3.661173236 3.014858740 0.1620643
#> [10,] 0.224584997 1.334680726 0.0585472

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.45017583 5.08543779 1.82493737
#>  [2,] 0.02950188 0.02950188 0.02950188
#>  [3,] 1.34054819 2.58714678 2.02237162
#>  [4,] 1.18202632 3.67379427 2.14307531
#>  [5,] 1.50810295 0.78652670 1.78155132
#>  [6,] 1.41545935 0.50504565 1.41545935
#>  [7,] 0.29618885 0.59367475 0.44836406
#>  [8,] 4.95182874 1.92963006 0.13799028
#>  [9,] 7.52646784 0.12655587 0.42890686
#> [10,] 0.76854449 0.06840268 0.76854449
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.69877661 0.69877661 0.69877661
#>  [2,] 0.01608014 0.01608014 0.01608014
#>  [3,] 0.18514814 0.18514814 0.18514814
#>  [4,] 2.23324086 2.23324086 2.23324086
#>  [5,] 0.28686215 0.28686215 0.28686215
#>  [6,] 0.01529286 0.01529286 0.01529286
#>  [7,] 0.30706731 0.30706731 0.30706731
#>  [8,] 0.93784697 0.93784697 0.93784697
#>  [9,] 0.92029016 0.92029016 0.92029016
#> [10,] 0.32084866 0.32084866 0.32084866
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]       [,2]        [,3]
#>  [1,] 1.98486804 1.43657072 0.266916749
#>  [2,] 0.06598884 0.26584417 0.046679191
#>  [3,] 0.21903809 0.39309641 0.784887882
#>  [4,] 0.02691825 3.50922132 0.036747293
#>  [5,] 1.53714891 0.09702195 1.297046320
#>  [6,] 0.95924369 0.16317648 0.587788372
#>  [7,] 2.96610600 1.13193486 0.144215715
#>  [8,] 1.47133537 0.55043022 0.006233884
#>  [9,] 0.27345913 1.24743288 0.151089952
#> [10,] 0.18280553 3.76289949 0.911834173

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,]  2.2154980 0.68760420 0.8814356
#>  [2,]  6.0174465 2.54796775 2.1120142
#>  [3,] 11.1184402 1.60672941 1.3260531
#>  [4,]  0.7559565 1.60662305 0.8587198
#>  [5,]  4.0006957 1.08378013 0.7079286
#>  [6,]  1.4941736 0.22499923 0.1021902
#>  [7,]  0.7363976 0.08070248 0.3341892
#>  [8,]  0.5350414 0.29640372 3.2398481
#>  [9,]  0.5625859 1.89471391 6.6094324
#> [10,]  0.4964589 0.08919104 0.4964589
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0544952 1.0544952 1.0544952
#>  [2,] 0.3658314 0.3658314 0.3658314
#>  [3,] 1.1791666 1.1791666 1.1791666
#>  [4,] 1.0239948 1.0239948 1.0239948
#>  [5,] 0.5092481 0.5092481 0.5092481
#>  [6,] 1.3768438 1.3768438 1.3768438
#>  [7,] 1.1146624 1.1146624 1.1146624
#>  [8,] 0.7899009 0.7899009 0.7899009
#>  [9,] 0.2374284 0.2374284 0.2374284
#> [10,] 0.9012778 0.9012778 0.9012778
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>               [,1]       [,2]       [,3]
#>  [1,] 1.9063343181 0.52270417 2.01167578
#>  [2,] 0.5418303045 0.36779300 0.20412827
#>  [3,] 0.0007608463 0.14582519 1.35161705
#>  [4,] 0.0170852495 0.08269218 0.29780517
#>  [5,] 1.9014385472 0.13680691 1.34514240
#>  [6,] 2.4243259447 1.37979478 0.06221833
#>  [7,] 2.4733133967 3.53585455 0.23345730
#>  [8,] 2.7540246266 2.11446525 1.84729690
#>  [9,] 0.0059734417 0.09345217 2.46071264
#> [10,] 0.5833893972 1.27827076 1.06394232

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 3.22362459 5.7807463 0.28482019
#>  [2,] 3.41805355 0.1370594 0.06708591
#>  [3,] 3.23522779 0.5135232 4.06706824
#>  [4,] 4.86604458 4.1665613 2.11886935
#>  [5,] 1.86051733 2.0697019 5.89990098
#>  [6,] 1.40100240 1.4010024 1.40100240
#>  [7,] 0.05229757 1.5577200 1.38306573
#>  [8,] 1.73618639 1.0001317 1.73618639
#>  [9,] 0.41875845 0.4187584 0.41875845
#> [10,] 0.16263124 0.1626312 0.16263124
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60006163 0.60006163 0.60006163
#>  [2,] 1.62490128 1.62490128 1.62490128
#>  [3,] 0.02776522 0.02776522 0.02776522
#>  [4,] 0.33925547 0.33925547 0.33925547
#>  [5,] 0.39512812 0.39512812 0.39512812
#>  [6,] 1.13014006 1.13014006 1.13014006
#>  [7,] 2.82331875 2.82331875 2.82331875
#>  [8,] 0.65618879 0.65618879 0.65618879
#>  [9,] 1.02255482 1.02255482 1.02255482
#> [10,] 0.35474805 0.35474805 0.35474805
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03403085 0.74316051 0.08258863
#>  [2,] 1.86733858 0.51086725 0.84494400
#>  [3,] 0.94013026 0.45646801 0.70964841
#>  [4,] 0.62891785 0.05610156 0.12429769
#>  [5,] 0.36599925 1.23558079 0.72310346
#>  [6,] 0.21694355 1.09906701 0.08753187
#>  [7,] 0.33053075 1.51474978 0.05633423
#>  [8,] 0.43903697 0.65953736 1.30631574
#>  [9,] 1.62619426 0.87870250 0.87602182
#> [10,] 0.15830442 0.33432316 0.57964234

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.81191420 0.61150506 0.69392611
#>  [2,] 1.64970514 2.00613793 2.20493262
#>  [3,] 1.26703766 1.26703766 1.03167908
#>  [4,] 1.95302680 1.84897050 1.95302680
#>  [5,] 3.07168345 1.08202087 3.96750054
#>  [6,] 1.22296639 0.19547602 2.18623042
#>  [7,] 0.06181781 0.06181781 0.06181781
#>  [8,] 0.18489818 0.21736523 0.32677327
#>  [9,] 0.86788615 0.54882824 0.87849690
#> [10,] 0.32155929 5.62292196 0.78032080
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.05392597 1.05392597 1.05392597
#>  [2,] 0.08979148 0.08979148 0.08979148
#>  [3,] 2.37053844 2.37053844 2.37053844
#>  [4,] 0.11787737 0.11787737 0.11787737
#>  [5,] 0.23842831 0.23842831 0.23842831
#>  [6,] 0.95993059 0.95993059 0.95993059
#>  [7,] 0.52727718 0.52727718 0.52727718
#>  [8,] 0.67915695 0.67915695 0.67915695
#>  [9,] 2.71745054 2.71745054 2.71745054
#> [10,] 0.03722171 0.03722171 0.03722171
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]       [,2]        [,3]
#>  [1,] 0.9020570 0.64655138 1.676155474
#>  [2,] 0.4237699 0.63633593 2.599481119
#>  [3,] 0.3215959 0.27077771 0.945493059
#>  [4,] 0.1025318 0.03754799 0.006463063
#>  [5,] 0.3553571 0.51060029 0.044640027
#>  [6,] 1.4155333 0.51969844 0.838474471
#>  [7,] 2.2304988 0.29236050 0.078986240
#>  [8,] 2.2512597 0.24117172 1.515080627
#>  [9,] 0.5107427 0.93454489 1.320330489
#> [10,] 1.5423036 0.95666101 1.438769352

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.4000689 0.87423431 0.4380265
#>  [2,] 0.6438385 3.58932307 2.0885857
#>  [3,] 1.4508271 4.04801461 3.5576239
#>  [4,] 1.3998989 1.52517100 1.5251710
#>  [5,] 1.9940607 4.70430370 4.7043037
#>  [6,] 4.2797315 1.40391078 0.8266957
#>  [7,] 0.6760121 0.67601208 0.6760121
#>  [8,] 4.2125523 0.01285755 1.7593842
#>  [9,] 3.2997435 3.62894958 8.1320322
#> [10,] 1.1846579 1.72522076 3.4884723
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.79424803 0.79424803 0.79424803
#>  [2,] 1.60904976 1.60904976 1.60904976
#>  [3,] 0.27902351 0.27902351 0.27902351
#>  [4,] 0.16089903 0.16089903 0.16089903
#>  [5,] 0.27649828 0.27649828 0.27649828
#>  [6,] 1.01302509 1.01302509 1.01302509
#>  [7,] 0.41123187 0.41123187 0.41123187
#>  [8,] 1.54679751 1.54679751 1.54679751
#>  [9,] 1.69329378 1.69329378 1.69329378
#> [10,] 0.02104718 0.02104718 0.02104718
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]        [,2]      [,3]
#>  [1,] 0.26229678 1.394108825 1.2387640
#>  [2,] 0.33405344 0.893898221 0.0699110
#>  [3,] 0.01452577 1.588819477 1.1805432
#>  [4,] 0.66801175 0.303057169 0.2784330
#>  [5,] 0.34409818 0.110067321 0.3729206
#>  [6,] 0.67285230 0.536867921 0.8278428
#>  [7,] 1.16568868 0.067512717 0.2158858
#>  [8,] 1.00541830 0.666026983 1.0532493
#>  [9,] 0.03292002 0.061975273 1.1475601
#> [10,] 0.44220526 0.008760339 0.3125543

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.03677504 0.2071274 0.2071274
#>  [2,] 0.72801857 0.5929393 0.7280186
#>  [3,] 0.87297828 0.2462049 1.8575600
#>  [4,] 0.06600693 0.1627355 0.4645862
#>  [5,] 0.68914809 1.0461098 0.6891481
#>  [6,] 0.49400311 0.6694169 2.3735586
#>  [7,] 1.17495601 0.0804637 1.3199833
#>  [8,] 1.84281858 1.2263073 0.4053579
#>  [9,] 0.49780987 2.4270070 0.3793778
#> [10,] 1.12932834 1.9287857 1.1293283
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.03861906 1.5521151 2.12404359
#>  [2,] 0.58268369 0.3928541 0.39285407
#>  [3,] 0.88975687 0.9373386 0.27657578
#>  [4,] 1.32447781 1.7789027 0.37354803
#>  [5,] 0.35805023 0.3580502 0.18755395
#>  [6,] 1.50185688 0.6952363 0.69523631
#>  [7,] 0.77513296 1.0086130 0.27569162
#>  [8,] 0.49956753 0.5398419 0.04717786
#>  [9,] 0.45796383 0.3293780 0.61051450
#> [10,] 0.03834530 0.7829338 0.78293380
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4750116 0.1820215 0.4871160
#>  [2,] 1.9941447 0.2016480 0.4404778
#>  [3,] 0.6748159 0.6748159 3.3671806
#>  [4,] 1.0813147 0.4347707 0.3432397
#>  [5,] 2.3321349 0.2317952 0.5650854
#>  [6,] 0.7004021 0.1898738 0.7004021
#>  [7,] 1.0538283 0.3286439 1.4703057
#>  [8,] 2.2215978 2.6448750 1.6480544
#>  [9,] 1.0862659 1.0924898 1.0924898
#> [10,] 0.9447962 1.6431981 1.3177961
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12198108 1.68352357 0.96402632
#>  [2,] 0.36736257 2.57764953 0.75322877
#>  [3,] 0.22350088 0.11725434 0.11725434
#>  [4,] 0.86896301 1.35555569 1.23397941
#>  [5,] 0.50946318 0.50946318 0.50946318
#>  [6,] 0.41946169 0.43286646 1.83232273
#>  [7,] 1.69105586 2.70135810 0.20701186
#>  [8,] 0.01165992 0.01165992 0.01165992
#>  [9,] 1.04493575 0.09909770 1.05528676
#> [10,] 0.61356899 0.61356899 0.25029599
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.35357997 0.353579974 0.35357997
#>  [2,] 0.16262289 1.262071653 0.16262289
#>  [3,] 0.06935676 0.069356760 0.23613053
#>  [4,] 0.13391843 0.133918427 0.13391843
#>  [5,] 0.55666889 0.556668888 0.55666889
#>  [6,] 0.54806676 0.570195247 1.65377424
#>  [7,] 0.30862377 0.343212062 0.85192079
#>  [8,] 0.57013353 0.009471879 0.22070399
#>  [9,] 0.48537384 0.485373843 0.01690307
#> [10,] 0.03373814 1.259426591 1.37020335

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.35571481 2.3557148 2.3557148
#>  [2,] 0.52799747 2.0395058 5.5241986
#>  [3,] 0.02715018 1.1583055 0.1750333
#>  [4,] 5.54670223 0.4274940 5.4079969
#>  [5,] 2.22164261 2.0670158 2.2216426
#>  [6,] 0.18983054 0.1375339 3.2829982
#>  [7,] 2.03998175 0.6725197 0.1234812
#>  [8,] 0.82465997 1.2854386 2.7290474
#>  [9,] 4.45748428 1.8445516 0.1422299
#> [10,] 0.68033200 0.3714602 1.4018786
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>             [,1]        [,2]      [,3]
#>  [1,] 0.10946297 0.667348201 2.2628970
#>  [2,] 0.87985409 1.176782657 1.3985587
#>  [3,] 5.89200687 3.457839015 0.4709756
#>  [4,] 0.12208534 0.122085344 0.1220853
#>  [5,] 0.06006096 1.708127414 2.5919024
#>  [6,] 1.20794203 7.125320622 0.2331707
#>  [7,] 0.14170063 0.141700634 0.1417006
#>  [8,] 1.15305942 0.008123914 2.2107019
#>  [9,] 0.11852357 1.065042148 0.3978543
#> [10,] 2.50341993 0.471619951 1.0080663
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 5.04054274 3.6220996 1.19315798
#>  [2,] 3.36725799 2.5215795 1.05916238
#>  [3,] 2.82135519 3.9955687 1.08134581
#>  [4,] 2.08899307 0.6779082 0.62827913
#>  [5,] 5.19370430 1.4075417 2.73838929
#>  [6,] 0.06225482 0.6193466 0.01190762
#>  [7,] 1.52033762 1.8030072 3.66868894
#>  [8,] 2.82650110 0.3554843 4.06882397
#>  [9,] 0.71728775 1.5226255 4.40523410
#> [10,] 5.35334881 0.9420711 0.87473370
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.30564169 1.29502147 0.77804456
#>  [2,] 0.04788458 0.04788458 0.04788458
#>  [3,] 1.71164477 0.51924050 0.38583430
#>  [4,] 0.13970901 0.33741944 0.07371916
#>  [5,] 6.76216309 0.38163460 1.95458630
#>  [6,] 0.58399147 0.58399147 0.21841958
#>  [7,] 0.12844929 1.12992449 1.09823452
#>  [8,] 2.20404944 2.49683555 0.67473663
#>  [9,] 0.27577606 0.35876938 1.83797218
#> [10,] 0.72905201 0.03523142 2.22404744
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7076887 1.53486264 1.2572367
#>  [2,] 0.8929210 2.94277558 2.3696069
#>  [3,] 3.7653247 0.43574607 2.6017405
#>  [4,] 4.4462112 1.12262512 0.6639328
#>  [5,] 1.5493819 0.01943693 1.1839329
#>  [6,] 2.4598810 1.61484318 1.5287209
#>  [7,] 0.1629147 1.84104941 1.0530486
#>  [8,] 2.3731334 1.15007444 4.7758274
#>  [9,] 1.2340321 1.45578125 1.6955041
#> [10,] 0.8664929 0.47707516 0.8664929

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.10929875 0.10929875 0.1092987
#>  [2,] 1.08547130 1.78994604 1.0854713
#>  [3,] 0.07769039 1.48202895 0.8837241
#>  [4,] 0.42632391 0.09774453 1.4083727
#>  [5,] 0.06128176 0.43196708 0.4319671
#>  [6,] 0.11700263 0.11700263 0.1170026
#>  [7,] 0.98351476 1.29075264 0.9835148
#>  [8,] 0.28156854 0.28156854 0.2815685
#>  [9,] 0.49615208 0.18656446 0.8562074
#> [10,] 0.10086132 0.39801244 0.2551755
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3143823 0.6890157 0.08051804
#>  [2,] 0.1148554 0.3017754 0.30177541
#>  [3,] 0.4172451 0.4172451 0.11730706
#>  [4,] 0.7347128 0.7347128 0.73471277
#>  [5,] 0.4503940 0.1653738 0.46008503
#>  [6,] 0.2733353 0.2733353 0.27333527
#>  [7,] 0.3101127 0.3145853 0.15811496
#>  [8,] 0.3026004 0.3026004 0.30260039
#>  [9,] 0.6964499 0.4248364 0.69644986
#> [10,] 1.2087718 0.4638149 1.79166435
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9678168 0.02094485 0.9678168
#>  [2,] 0.8250629 1.23213418 0.1933162
#>  [3,] 0.1676144 0.16761438 0.1676144
#>  [4,] 0.5728697 0.57286968 0.5728697
#>  [5,] 0.7398840 0.73988398 0.3988762
#>  [6,] 0.4760008 0.47600081 0.1129219
#>  [7,] 0.1258459 0.12584591 0.1258459
#>  [8,] 0.1007326 1.67848321 0.1007326
#>  [9,] 0.4907638 0.79412804 0.4907638
#> [10,] 0.2199387 0.21993868 0.2199387
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 1.19053996 0.003048023 1.19053996
#>  [2,] 0.04048565 0.040485646 0.04048565
#>  [3,] 0.17914915 0.179149147 0.14487823
#>  [4,] 0.24945958 0.249459583 0.24945958
#>  [5,] 0.83013807 1.007621445 0.83013807
#>  [6,] 0.08093811 0.068194600 0.08093811
#>  [7,] 0.26832701 0.150686181 1.02089715
#>  [8,] 0.28171144 0.281711438 0.28171144
#>  [9,] 0.39279481 0.600274363 0.20753476
#> [10,] 0.37374399 0.373743992 0.37374399
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.008333051 0.1915545 0.3662792
#>  [2,] 0.978409613 0.8025397 0.3392906
#>  [3,] 0.642689698 0.7122916 0.6426897
#>  [4,] 0.082478072 0.2456947 0.2456947
#>  [5,] 0.555499036 0.5554990 0.5344852
#>  [6,] 0.181196265 0.1811963 0.1811963
#>  [7,] 0.315439982 0.3154400 0.3154400
#>  [8,] 0.951894515 1.0256394 1.0256394
#>  [9,] 0.905479492 0.1777552 0.9054795
#> [10,] 0.478201901 1.1507733 1.1507733

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4515894 0.45158938 0.26410947
#>  [2,] 0.9892620 0.35680697 0.52195685
#>  [3,] 0.1761819 0.71643408 0.38683716
#>  [4,] 0.6534684 0.24526299 0.05584853
#>  [5,] 1.2770412 0.38057851 0.07795208
#>  [6,] 0.2276945 0.22769450 0.22769450
#>  [7,] 0.3872656 0.38726556 0.38003825
#>  [8,] 0.6816465 0.15917593 0.37387957
#>  [9,] 0.2368176 0.02136835 0.54390914
#> [10,] 0.1043171 0.67854450 0.51299985
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.69385057 0.17094468 0.33549161
#>  [2,] 0.01032634 0.39810092 0.39810092
#>  [3,] 0.05538987 0.03359319 0.17459193
#>  [4,] 0.08333055 0.08333055 0.08333055
#>  [5,] 0.43947515 0.13033872 0.25894492
#>  [6,] 0.05000509 0.51258948 0.25682134
#>  [7,] 0.40742061 0.08005438 0.93000371
#>  [8,] 0.31165988 0.92856049 0.92856049
#>  [9,] 0.49090901 0.49090901 0.49090901
#> [10,] 0.44506212 0.38593064 0.70854535
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.255303230 0.25530323 1.22488834
#>  [2,] 0.442095590 0.91190833 1.90627127
#>  [3,] 0.449115350 0.44911535 0.44911535
#>  [4,] 0.064818628 0.08077094 0.06481863
#>  [5,] 0.997594764 0.30833091 0.55351681
#>  [6,] 0.003095544 0.33010367 0.33010367
#>  [7,] 0.750145948 0.75014595 0.75014595
#>  [8,] 0.114191867 0.11419187 0.11419187
#>  [9,] 0.246158747 0.24615875 0.32891728
#> [10,] 0.011789577 0.01178958 0.01178958
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32871712 0.03516347 0.32871712
#>  [2,] 0.04662698 0.04662698 0.04662698
#>  [3,] 0.34764718 0.34764718 0.65808277
#>  [4,] 0.69836217 0.16158517 0.71207255
#>  [5,] 0.27533238 0.15767728 0.27533238
#>  [6,] 1.32942109 0.46715403 0.70303174
#>  [7,] 0.03686898 0.03686898 0.03686898
#>  [8,] 0.08326702 0.67064612 0.08326702
#>  [9,] 0.03813711 0.03813711 0.03813711
#> [10,] 0.81990222 0.23936401 0.06874254

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.25510568 1.89303709 1.24873331
#>  [2,] 1.50476330 0.01389896 1.17606920
#>  [3,] 0.01451169 0.98825996 0.26712443
#>  [4,] 0.02976312 0.25954903 0.40113454
#>  [5,] 1.07875784 0.02951054 0.05306956
#>  [6,] 0.27760546 0.38402306 0.27760546
#>  [7,] 0.17949871 0.17949871 0.17949871
#>  [8,] 0.47345137 1.14336562 0.47345137
#>  [9,] 1.27241414 1.32479931 1.05078244
#> [10,] 0.06274734 0.06274734 0.06274734
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12939795 0.12939795 0.12939795
#>  [2,] 0.56781014 0.29587712 0.01491664
#>  [3,] 1.89391797 0.96228876 0.96228876
#>  [4,] 0.54408345 0.08374799 0.54408345
#>  [5,] 0.09390462 0.09390462 0.20263223
#>  [6,] 0.39580981 0.06251219 0.39580981
#>  [7,] 0.10322420 0.38841662 0.16015334
#>  [8,] 0.15742842 0.10315966 0.03826262
#>  [9,] 0.13868190 0.13868190 0.13868190
#> [10,] 0.38745606 0.75235002 0.75235002
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.324848790 0.09212486 0.11145343
#>  [2,] 0.047546540 0.47173502 0.34650895
#>  [3,] 0.016367379 0.14556883 0.01636738
#>  [4,] 0.331239192 1.58840155 0.56676265
#>  [5,] 0.141099605 0.44027479 0.70669353
#>  [6,] 0.258505614 0.19729379 0.57552283
#>  [7,] 0.011161501 0.01116150 0.01116150
#>  [8,] 2.833698745 0.04422954 0.06245329
#>  [9,] 0.792109871 0.42270659 0.19202973
#> [10,] 0.006339606 0.31653945 0.31653945
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.10552958 0.10552958 0.03764788
#>  [2,] 0.03134312 0.03134312 0.03134312
#>  [3,] 0.39041168 0.06291383 0.12448769
#>  [4,] 0.31165976 0.05998157 0.59037886
#>  [5,] 0.09289036 0.80226952 0.09289036
#>  [6,] 0.13358920 0.69694101 0.14158294
#>  [7,] 0.05338839 0.05338839 0.05338839
#>  [8,] 0.56520959 0.11859900 0.03656910
#>  [9,] 1.01116410 0.31824616 0.33692257
#> [10,] 0.67471358 0.35358227 0.30159577

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58771535 0.17701124 0.17701124
#>  [2,] 0.08080998 0.08080998 0.08080998
#>  [3,] 0.45931111 0.40258300 0.40258300
#>  [4,] 0.06355540 0.06355540 0.06355540
#>  [5,] 0.95121280 1.11441149 0.33778517
#>  [6,] 1.20264593 0.05660056 0.10795602
#>  [7,] 0.61551675 0.43788702 0.17511737
#>  [8,] 0.39781891 0.39781891 0.65317295
#>  [9,] 0.51119218 0.30201169 0.10395392
#> [10,] 1.58913189 0.17852590 0.89137460
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36339825 0.18059798 0.36339825
#>  [2,] 0.06158708 0.35613812 0.06158708
#>  [3,] 0.02153235 0.23899977 0.23899977
#>  [4,] 0.65956865 0.65956865 0.65956865
#>  [5,] 0.09608233 0.09608233 0.09608233
#>  [6,] 0.33159063 0.71111117 0.48966383
#>  [7,] 0.58629036 0.07001716 0.58629036
#>  [8,] 0.33731273 0.65165768 0.33731273
#>  [9,] 0.04268445 0.04268445 0.04268445
#> [10,] 0.07147596 0.90225192 2.29878209
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.61362639 0.64504216 0.29481286
#>  [2,] 0.10145696 0.09546035 0.10145696
#>  [3,] 0.11636121 0.24883307 0.11636121
#>  [4,] 0.07906811 0.58545437 0.23772949
#>  [5,] 0.21678946 0.06028082 0.21678946
#>  [6,] 1.27352559 0.74371239 0.81112341
#>  [7,] 0.03916826 0.24058096 0.03916826
#>  [8,] 0.26857973 0.02099993 0.42511580
#>  [9,] 0.81596008 0.81596008 0.22553593
#> [10,] 0.06363016 0.06363016 0.06363016
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.71221239 0.71221239 0.56535740
#>  [2,] 0.05079635 0.05079635 0.05079635
#>  [3,] 0.25257254 1.39778484 0.62726456
#>  [4,] 0.84910340 0.61090287 0.43153043
#>  [5,] 0.25920988 0.32376974 0.26364078
#>  [6,] 0.71578556 0.15768740 0.71866665
#>  [7,] 0.23691494 0.23691494 0.23691494
#>  [8,] 0.19905891 0.10407547 0.10407547
#>  [9,] 0.10507731 0.19248575 0.19248575
#> [10,] 0.15924692 1.58543000 0.08176109