Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4436330 0.4436330 0.44363299
#>  [2,] 0.7026227 6.1147585 0.45097367
#>  [3,] 3.8452707 3.8452707 0.15074575
#>  [4,] 0.5479648 0.9725060 0.97250598
#>  [5,] 2.8230827 5.0507046 1.09837951
#>  [6,] 0.2570811 3.5963134 2.06495805
#>  [7,] 1.7351765 0.5617798 3.79120619
#>  [8,] 2.6645529 5.2725923 0.09895101
#>  [9,] 1.1130422 0.4629231 0.81603455
#> [10,] 0.6682293 0.3301310 2.14051750
## comonotone
rpextmo(10, 3, a = 1)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.316198767 0.316198767 0.316198767
#>  [2,] 0.408281612 0.408281612 0.408281612
#>  [3,] 0.745915119 0.745915119 0.745915119
#>  [4,] 0.004305083 0.004305083 0.004305083
#>  [5,] 0.566194247 0.566194247 0.566194247
#>  [6,] 0.116951800 0.116951800 0.116951800
#>  [7,] 0.035554003 0.035554003 0.035554003
#>  [8,] 2.324764541 2.324764541 2.324764541
#>  [9,] 1.400843123 1.400843123 1.400843123
#> [10,] 2.741532316 2.741532316 2.741532316
## independence
rpextmo(10, 3, b = 1)
#>            [,1]       [,2]      [,3]
#>  [1,] 3.3243085 0.16907203 7.6655221
#>  [2,] 0.2648396 3.39901490 0.4044777
#>  [3,] 0.4424226 0.04084144 3.9615947
#>  [4,] 0.2766354 4.07546525 0.5352220
#>  [5,] 0.9589666 1.45449355 0.8188510
#>  [6,] 0.9610609 0.13949574 0.8452050
#>  [7,] 0.2866990 0.29860146 0.2203041
#>  [8,] 0.2285130 0.40904304 0.5068410
#>  [9,] 1.3028879 1.30640962 1.5276764
#> [10,] 0.3015682 1.04738159 1.0511450

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.20999771 2.76815762 0.7517011
#>  [2,] 0.61075214 0.61075214 0.6107521
#>  [3,] 2.31677664 0.75579481 5.2817210
#>  [4,] 0.06515426 0.33563638 1.9549983
#>  [5,] 1.39141353 1.58002838 0.9925677
#>  [6,] 2.70268815 2.50087266 0.5985558
#>  [7,] 2.30728759 0.18720325 1.3675064
#>  [8,] 3.40287349 0.96204875 0.2087884
#>  [9,] 1.18595743 1.18595743 0.3039896
#> [10,] 0.69654940 0.01134636 0.1528119
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03521921 0.03521921 0.03521921
#>  [2,] 1.87502690 1.87502690 1.87502690
#>  [3,] 1.35831329 1.35831329 1.35831329
#>  [4,] 1.40703882 1.40703882 1.40703882
#>  [5,] 0.07934278 0.07934278 0.07934278
#>  [6,] 0.07933640 0.07933640 0.07933640
#>  [7,] 0.38560294 0.38560294 0.38560294
#>  [8,] 0.33926294 0.33926294 0.33926294
#>  [9,] 1.79975444 1.79975444 1.79975444
#> [10,] 0.98457321 0.98457321 0.98457321
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]      [,2]         [,3]
#>  [1,] 3.91786090 0.3412708 0.0007120096
#>  [2,] 0.02227263 2.0652062 0.8529612021
#>  [3,] 1.14778306 1.1773130 0.8435486620
#>  [4,] 2.25489054 0.6925519 0.0317977830
#>  [5,] 2.19795839 0.1107033 0.5389296795
#>  [6,] 1.33520463 0.6764825 0.2304078434
#>  [7,] 0.61528823 0.8905715 0.3717594342
#>  [8,] 0.11117799 1.6834684 0.2359771351
#>  [9,] 1.32227618 0.9260536 0.6015491229
#> [10,] 0.24879259 2.2552160 0.2347238613

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.36435839 2.9951016 0.4505196
#>  [2,] 5.68772086 5.6877209 0.7969359
#>  [3,] 3.75570321 1.1837373 1.1673068
#>  [4,] 0.01764057 2.7257733 2.7257733
#>  [5,] 0.21512443 0.4382712 0.4382712
#>  [6,] 0.41804058 0.3838232 0.4180406
#>  [7,] 2.39284667 0.6385884 2.3928467
#>  [8,] 0.20043448 2.3258912 0.8401074
#>  [9,] 0.59648100 1.0333894 0.5541924
#> [10,] 4.83083612 1.9629153 0.6482105
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7752863 0.7752863 0.7752863
#>  [2,] 0.3548441 0.3548441 0.3548441
#>  [3,] 2.4007863 2.4007863 2.4007863
#>  [4,] 0.5488947 0.5488947 0.5488947
#>  [5,] 0.9334031 0.9334031 0.9334031
#>  [6,] 0.9081062 0.9081062 0.9081062
#>  [7,] 0.2643348 0.2643348 0.2643348
#>  [8,] 0.2134350 0.2134350 0.2134350
#>  [9,] 0.1486578 0.1486578 0.1486578
#> [10,] 0.2186276 0.2186276 0.2186276
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]      [,2]        [,3]
#>  [1,] 0.2331992 1.9352783 1.528204227
#>  [2,] 1.1668741 0.4465768 0.002544635
#>  [3,] 0.5549020 0.9140001 1.014443456
#>  [4,] 0.1023532 0.6402820 0.372707478
#>  [5,] 1.5066504 0.2827681 0.447486992
#>  [6,] 0.9477815 0.9601798 0.286669954
#>  [7,] 0.1792507 0.8044327 1.693008203
#>  [8,] 0.5261610 1.3970494 1.662748255
#>  [9,] 0.4137492 0.4064082 3.149700969
#> [10,] 1.9971909 0.4917908 1.832688354

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.32447752 0.49233715 0.91512839
#>  [2,] 3.00026347 4.50433440 0.37703578
#>  [3,] 1.18480715 0.33828902 0.16925471
#>  [4,] 0.16395120 0.03534403 1.86151990
#>  [5,] 1.28499324 0.96671079 1.90463117
#>  [6,] 1.07621923 3.68247169 5.94696582
#>  [7,] 3.95239849 1.45387708 1.20685167
#>  [8,] 0.08991516 0.08991516 0.08991516
#>  [9,] 0.34493197 0.90547994 0.90547994
#> [10,] 1.52135888 0.53101532 1.75028675
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0622564 1.0622564 1.0622564
#>  [2,] 0.2928866 0.2928866 0.2928866
#>  [3,] 0.5588151 0.5588151 0.5588151
#>  [4,] 1.3522801 1.3522801 1.3522801
#>  [5,] 0.1262882 0.1262882 0.1262882
#>  [6,] 4.8481627 4.8481627 4.8481627
#>  [7,] 1.3552571 1.3552571 1.3552571
#>  [8,] 0.2387467 0.2387467 0.2387467
#>  [9,] 1.0680450 1.0680450 1.0680450
#> [10,] 0.1574200 0.1574200 0.1574200
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.49582837 2.8343139 0.16224461
#>  [2,] 0.12909921 2.2752991 0.29076866
#>  [3,] 0.91219610 0.8753445 1.06877232
#>  [4,] 0.10482954 0.4380589 0.05110368
#>  [5,] 1.32489809 0.3379825 0.61755002
#>  [6,] 0.69134273 0.3794173 0.92694612
#>  [7,] 1.61011063 0.7050509 2.66123170
#>  [8,] 0.04034584 1.3941848 0.89828562
#>  [9,] 0.70271719 0.6048012 0.23283747
#> [10,] 0.73544808 1.1387046 1.30617834

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.9962354 0.72683696 0.4605551
#>  [2,] 0.5275206 3.56637588 0.1700604
#>  [3,] 0.1663724 0.14677289 0.1663724
#>  [4,] 0.2664286 1.39048608 0.4251755
#>  [5,] 0.7892915 0.78929154 0.7892915
#>  [6,] 0.9823576 0.98235757 0.2593614
#>  [7,] 3.6312111 0.03571082 0.8235589
#>  [8,] 0.6970200 1.05774014 0.1331496
#>  [9,] 0.7186093 6.18639160 3.4320276
#> [10,] 0.5444434 0.72638769 0.8564415
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.44088340 1.44088340 1.44088340
#>  [2,] 0.01453715 0.01453715 0.01453715
#>  [3,] 1.10878633 1.10878633 1.10878633
#>  [4,] 4.04284283 4.04284283 4.04284283
#>  [5,] 2.34932053 2.34932053 2.34932053
#>  [6,] 1.17399471 1.17399471 1.17399471
#>  [7,] 3.03836786 3.03836786 3.03836786
#>  [8,] 0.22208199 0.22208199 0.22208199
#>  [9,] 0.29049869 0.29049869 0.29049869
#> [10,] 0.07042619 0.07042619 0.07042619
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.13601097 0.8479169 0.23770497
#>  [2,] 0.10012152 0.4390403 0.46897842
#>  [3,] 0.29400983 0.4429200 1.17799802
#>  [4,] 0.18047645 0.3622384 0.70631365
#>  [5,] 0.80564039 0.7474875 0.45786386
#>  [6,] 0.38414769 0.2105582 0.03838763
#>  [7,] 0.48302507 1.7085619 5.61148591
#>  [8,] 0.96460123 0.5263692 0.31315441
#>  [9,] 1.38374279 0.1732873 1.77122978
#> [10,] 0.02299565 0.1012953 0.53484247

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1261511 3.4022511 1.16867909
#>  [2,] 0.4148563 1.5533118 0.01129877
#>  [3,] 0.5716768 1.7514515 2.70875771
#>  [4,] 0.3673374 0.3673374 0.34768987
#>  [5,] 0.0119308 1.1953759 0.90666891
#>  [6,] 2.2150635 4.5911875 1.97675651
#>  [7,] 0.8334718 0.8334718 0.33034174
#>  [8,] 0.1467077 3.2609834 3.53483150
#>  [9,] 0.5341901 0.5341901 0.53419006
#> [10,] 0.2130606 0.1148946 1.91392669
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.006710635 0.006710635 0.006710635
#>  [2,] 0.780909304 0.780909304 0.780909304
#>  [3,] 0.335212812 0.335212812 0.335212812
#>  [4,] 0.304949461 0.304949461 0.304949461
#>  [5,] 0.266591499 0.266591499 0.266591499
#>  [6,] 1.485877667 1.485877667 1.485877667
#>  [7,] 1.002063869 1.002063869 1.002063869
#>  [8,] 0.164877202 0.164877202 0.164877202
#>  [9,] 0.333738651 0.333738651 0.333738651
#> [10,] 1.608217368 1.608217368 1.608217368
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]       [,3]
#>  [1,] 2.39154671 0.7877437 0.07144076
#>  [2,] 1.16211073 0.3372385 3.15631065
#>  [3,] 0.19228815 0.6888885 0.12108943
#>  [4,] 0.10473276 1.5812183 2.53142376
#>  [5,] 0.36491537 0.4128179 0.02068107
#>  [6,] 1.42869075 0.8198960 2.79640020
#>  [7,] 2.31358404 0.7022951 0.41910646
#>  [8,] 1.51774056 1.8189534 2.17631595
#>  [9,] 0.54930549 1.1714957 0.80474275
#> [10,] 0.07549311 0.9169916 1.62858254

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6415668 0.4832145 1.0792993
#>  [2,] 3.0370097 1.8030019 0.1130871
#>  [3,] 1.0096866 0.9786879 1.0682745
#>  [4,] 0.0590756 0.4151362 0.2404053
#>  [5,] 1.1340605 0.9814833 0.3674035
#>  [6,] 1.6240932 1.7077037 0.7972225
#>  [7,] 0.4724147 0.4081888 2.2344678
#>  [8,] 0.8860650 1.2232962 0.7377866
#>  [9,] 0.1643602 2.1586800 0.1810002
#> [10,] 4.2503124 0.4381797 4.2503124
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.16443961 0.16443961 0.16443961
#>  [2,] 0.36482445 0.36482445 0.36482445
#>  [3,] 0.68988468 0.68988468 0.68988468
#>  [4,] 0.45349572 0.45349572 0.45349572
#>  [5,] 2.39038518 2.39038518 2.39038518
#>  [6,] 0.94371904 0.94371904 0.94371904
#>  [7,] 0.58246354 0.58246354 0.58246354
#>  [8,] 0.13288383 0.13288383 0.13288383
#>  [9,] 0.02865677 0.02865677 0.02865677
#> [10,] 0.25548390 0.25548390 0.25548390
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]        [,2]      [,3]
#>  [1,] 1.8088517 1.287784094 2.9877535
#>  [2,] 0.8660239 0.684806599 0.4765674
#>  [3,] 0.7304466 0.358406152 0.6966316
#>  [4,] 0.5972656 0.225978053 0.3712556
#>  [5,] 1.0990130 0.008901260 2.1109778
#>  [6,] 0.3397509 0.006502958 0.2322229
#>  [7,] 2.0969398 0.427230106 1.9713766
#>  [8,] 0.4824548 0.195293587 0.1399989
#>  [9,] 0.9737603 0.691830908 0.3976258
#> [10,] 0.1241129 0.276098962 1.3444389

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.38494037 3.8120541 0.6367330
#>  [2,] 0.04836261 2.1610546 1.8312630
#>  [3,] 1.12342130 0.1932952 3.4333297
#>  [4,] 2.95415330 0.3191493 2.8326424
#>  [5,] 0.09546367 1.7045070 0.9197459
#>  [6,] 2.03823801 0.3281157 1.8796399
#>  [7,] 4.98660798 2.4429320 3.0320996
#>  [8,] 1.23821092 3.0872520 0.9156089
#>  [9,] 1.66066672 0.6776127 1.9983946
#> [10,] 0.94662444 0.9019869 0.9466244
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.676861105 0.676861105 0.676861105
#>  [2,] 0.919542650 0.919542650 0.919542650
#>  [3,] 0.136726283 0.136726283 0.136726283
#>  [4,] 0.326535327 0.326535327 0.326535327
#>  [5,] 0.005129372 0.005129372 0.005129372
#>  [6,] 0.061867537 0.061867537 0.061867537
#>  [7,] 1.388771741 1.388771741 1.388771741
#>  [8,] 0.329811381 0.329811381 0.329811381
#>  [9,] 0.782922154 0.782922154 0.782922154
#> [10,] 3.737720370 3.737720370 3.737720370
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.24553974 0.8896423 3.80149313
#>  [2,] 0.59194212 1.3199096 3.13054794
#>  [3,] 2.86484907 0.3407042 0.05855175
#>  [4,] 2.46323264 0.3515825 0.66530680
#>  [5,] 1.13836889 0.4032406 0.08642938
#>  [6,] 2.11963230 1.1829797 0.89425083
#>  [7,] 0.95525653 0.4149853 2.25897161
#>  [8,] 4.09444537 1.2580351 1.13288699
#>  [9,] 0.70476217 0.0910627 0.29316147
#> [10,] 0.06761671 0.4040353 0.13424937

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>              [,1]       [,2]      [,3]
#>  [1,] 0.589786835 0.83186982 2.4935293
#>  [2,] 0.470805604 0.68015654 0.6801565
#>  [3,] 0.559242172 2.45381207 0.2041078
#>  [4,] 1.145132359 1.14513236 0.8765979
#>  [5,] 0.453285458 0.45328546 0.4532855
#>  [6,] 0.001456448 1.75889999 1.4867968
#>  [7,] 0.413855443 0.08470479 1.2458509
#>  [8,] 0.772854269 1.98708633 0.4939181
#>  [9,] 1.572507892 1.10235041 1.5725079
#> [10,] 1.789032722 1.25055889 3.5714736
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0187422 1.0187422 1.0187422
#>  [2,] 1.4745812 1.4745812 1.4745812
#>  [3,] 0.3695230 0.3695230 0.3695230
#>  [4,] 0.8113198 0.8113198 0.8113198
#>  [5,] 0.5226364 0.5226364 0.5226364
#>  [6,] 0.7484291 0.7484291 0.7484291
#>  [7,] 1.0961931 1.0961931 1.0961931
#>  [8,] 3.4888955 3.4888955 3.4888955
#>  [9,] 2.2292865 2.2292865 2.2292865
#> [10,] 0.2214398 0.2214398 0.2214398
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]        [,2]        [,3]
#>  [1,] 1.0799028 0.643426271 0.655499215
#>  [2,] 1.1874276 0.806913281 1.485061256
#>  [3,] 0.2813440 0.154619959 2.940737712
#>  [4,] 0.8116958 0.098087267 0.530548619
#>  [5,] 0.7048203 1.384861721 0.302005133
#>  [6,] 0.6580952 0.007704357 0.599780185
#>  [7,] 2.1755082 0.708259779 0.038060921
#>  [8,] 0.5112456 0.414533554 0.007626391
#>  [9,] 0.5653099 0.179218007 3.548504461
#> [10,] 0.6342061 0.219806881 1.026586099

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.06260036 0.53770337 1.06260036
#>  [2,] 1.51022722 1.51022722 1.51022722
#>  [3,] 0.32085360 4.30658705 2.85354025
#>  [4,] 0.43320379 3.60161963 0.18392953
#>  [5,] 0.38949169 0.54333065 2.24525561
#>  [6,] 4.33536016 0.91925102 2.33615704
#>  [7,] 0.38306168 1.41704453 1.41704453
#>  [8,] 4.07191418 0.11560947 0.74303318
#>  [9,] 0.04230696 0.04230696 0.04230696
#> [10,] 2.39593413 0.65915322 0.54280105
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.16287459 2.16287459 2.16287459
#>  [2,] 0.03938075 0.03938075 0.03938075
#>  [3,] 2.06591126 2.06591126 2.06591126
#>  [4,] 1.75613607 1.75613607 1.75613607
#>  [5,] 1.05670197 1.05670197 1.05670197
#>  [6,] 0.68908759 0.68908759 0.68908759
#>  [7,] 0.63309239 0.63309239 0.63309239
#>  [8,] 0.29935634 0.29935634 0.29935634
#>  [9,] 0.37820093 0.37820093 0.37820093
#> [10,] 2.31538521 2.31538521 2.31538521
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 2.41131766 1.4141176 3.7038292
#>  [2,] 0.38434860 3.9983052 1.8346538
#>  [3,] 0.04622170 1.8721944 0.2986078
#>  [4,] 0.01874549 0.4373887 1.1181563
#>  [5,] 0.34775474 0.4648853 0.1194525
#>  [6,] 0.50152098 1.1724141 0.4234454
#>  [7,] 1.06864473 1.0253372 0.6870938
#>  [8,] 0.10515372 0.6609282 0.8681967
#>  [9,] 0.25057454 1.6958712 0.6126759
#> [10,] 0.10650243 0.6151572 0.6589449

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38638645 1.42681945 0.38638645
#>  [2,] 0.32247614 1.28918107 0.68211781
#>  [3,] 0.24462784 0.06635312 0.06635312
#>  [4,] 0.31215808 0.31215808 2.16273376
#>  [5,] 0.07128950 0.42287718 0.07128950
#>  [6,] 0.01218405 0.01218405 0.01218405
#>  [7,] 0.15327985 1.34896297 0.15327985
#>  [8,] 0.60095776 0.60095776 0.39314468
#>  [9,] 0.14193509 0.15431026 0.02465432
#> [10,] 2.12805680 0.47249420 1.10622792
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.05625491 1.18635574 0.04488916
#>  [2,] 0.01364956 0.01364956 0.01364956
#>  [3,] 1.08754255 0.28244848 0.03546096
#>  [4,] 0.86131769 0.86131769 0.39902031
#>  [5,] 0.89372021 0.35354376 0.55768331
#>  [6,] 2.23028708 0.65679731 4.97471300
#>  [7,] 0.19364364 1.87204044 4.04566203
#>  [8,] 0.12416380 0.62490251 1.04835761
#>  [9,] 0.07825833 0.51833975 0.07825833
#> [10,] 0.13587752 0.13587752 0.13587752
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.79420590 0.06187882 0.27899708
#>  [2,] 0.40058695 0.96031985 0.61816673
#>  [3,] 0.02494898 0.02494898 0.02494898
#>  [4,] 0.84946941 0.32959448 0.36905395
#>  [5,] 0.96236910 0.42474106 2.05744347
#>  [6,] 0.80568137 0.62713851 0.80568137
#>  [7,] 0.97488394 1.02517422 1.14719477
#>  [8,] 1.10808518 0.39935043 0.25931712
#>  [9,] 0.57580805 0.57580805 0.57580805
#> [10,] 0.13376185 0.59522399 0.26350737
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>               [,1]        [,2]       [,3]
#>  [1,] 1.462464e+00 0.009121705 1.46246418
#>  [2,] 4.583003e-01 0.622485253 0.12149337
#>  [3,] 3.818770e-01 0.194014109 1.33847335
#>  [4,] 7.670725e-01 0.227620608 0.03412053
#>  [5,] 9.815479e-01 0.981547934 0.02203458
#>  [6,] 9.994701e-05 1.463470445 0.89511513
#>  [7,] 2.554377e-01 0.255437720 0.25543772
#>  [8,] 2.131693e-01 0.181784263 0.49507749
#>  [9,] 7.225280e-01 1.606600576 0.25966937
#> [10,] 5.498146e-01 1.511894273 2.46014387
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.77160932 0.26155566 0.8591426
#>  [2,] 0.82975642 0.04626757 1.1243152
#>  [3,] 0.81331879 0.41387455 0.7748773
#>  [4,] 0.03053381 0.11386975 0.6844836
#>  [5,] 1.82036619 1.34799646 0.8922603
#>  [6,] 0.26161223 0.49448601 1.7357016
#>  [7,] 0.42324130 0.42324130 0.4232413
#>  [8,] 0.15390053 0.15390053 0.1539005
#>  [9,] 0.66007820 0.66007820 0.6600782
#> [10,] 0.67334895 0.67334895 0.6733489

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.00468330 0.2647712 1.3337757
#>  [2,] 0.53290674 1.2502035 1.2502035
#>  [3,] 0.98370609 0.9837061 0.9837061
#>  [4,] 0.03985974 0.8963865 0.4713481
#>  [5,] 1.05595770 1.3891633 0.6906393
#>  [6,] 0.05913504 2.1872776 3.7560813
#>  [7,] 0.16345127 0.1634513 0.1634513
#>  [8,] 1.72266538 2.4812865 0.9483138
#>  [9,] 1.15027198 6.2708976 0.3205081
#> [10,] 0.16081754 0.3642266 0.6618648
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.12907018 0.14972430 1.12907018
#>  [2,] 0.78597406 0.74954180 1.43654595
#>  [3,] 2.48513729 1.21490325 2.48513729
#>  [4,] 0.28829386 1.20109983 1.20109983
#>  [5,] 0.09641755 0.09641755 0.09641755
#>  [6,] 0.55248840 1.80432422 1.43248710
#>  [7,] 2.73654318 7.75996785 1.86833804
#>  [8,] 0.30465351 0.30465351 0.30465351
#>  [9,] 1.75815566 0.47032056 0.32867776
#> [10,] 0.20713417 1.57708693 1.57708693
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 2.6952064 0.09491695 2.6952064
#>  [2,] 0.9404150 1.56158681 1.1650438
#>  [3,] 0.5055663 0.11274295 2.6119356
#>  [4,] 0.6132925 0.44217611 5.5975236
#>  [5,] 0.4179262 0.41792619 0.4179262
#>  [6,] 1.3359392 1.93130921 1.5107467
#>  [7,] 3.8747144 4.45607684 1.1117694
#>  [8,] 1.2461848 0.44218047 0.2518215
#>  [9,] 1.6203990 2.34091823 1.2436814
#> [10,] 1.9597624 5.73854192 4.0010880
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 2.83197995 2.1560224 0.3885951
#>  [2,] 1.66475685 0.1232252 1.3240377
#>  [3,] 0.39000231 0.3900023 0.3900023
#>  [4,] 1.24520616 0.9677582 7.3181444
#>  [5,] 0.02131884 0.4849267 0.4849267
#>  [6,] 1.09140325 0.2073838 0.3592377
#>  [7,] 1.56349870 2.7899517 1.6687928
#>  [8,] 0.59876906 0.5825452 0.1485107
#>  [9,] 3.27981608 4.3934842 1.1292738
#> [10,] 2.91943533 0.7629145 0.1171116
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.08856268 0.09110466 0.4937119
#>  [2,] 2.07456495 0.74081625 2.0745649
#>  [3,] 0.40403461 0.09139412 7.4843331
#>  [4,] 0.39807141 0.77842911 0.7784291
#>  [5,] 2.48421861 0.23886255 4.8992742
#>  [6,] 1.05813343 3.59619115 0.1962065
#>  [7,] 1.03005991 2.61130860 0.2199044
#>  [8,] 0.12089158 1.51835237 0.3946323
#>  [9,] 0.79020073 0.79020073 0.7902007
#> [10,] 0.03926702 2.05684344 2.0568434

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37950100 0.35648347 0.35648347
#>  [2,] 1.01803091 0.36583111 1.01803091
#>  [3,] 0.07847138 0.07847138 0.07847138
#>  [4,] 0.50261788 0.55389167 0.55389167
#>  [5,] 0.42820291 0.42820291 0.30586408
#>  [6,] 0.28155236 0.10801840 0.10801840
#>  [7,] 1.24237548 0.74642586 0.96289038
#>  [8,] 0.64297792 0.64297792 0.64297792
#>  [9,] 0.14359216 0.21507297 0.21507297
#> [10,] 0.05965905 0.15030797 0.74466987
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27689465 0.27689465 0.27689465
#>  [2,] 0.08641527 0.08641527 0.08641527
#>  [3,] 0.54121000 0.54121000 0.54121000
#>  [4,] 0.92322614 0.92322614 0.92322614
#>  [5,] 0.11114423 0.11114423 0.11114423
#>  [6,] 0.34957267 0.08719231 0.34957267
#>  [7,] 0.84628462 0.92895176 0.71637267
#>  [8,] 0.41716798 0.34431349 0.74016696
#>  [9,] 0.38710397 0.38710397 0.38710397
#> [10,] 0.49134615 0.25557804 0.51090246
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07096017 1.07096017 1.18570404
#>  [2,] 0.39615324 0.23893326 0.60387237
#>  [3,] 1.33134124 2.39343323 1.33134124
#>  [4,] 0.05155566 0.05155566 0.05155566
#>  [5,] 1.07137857 2.56867389 0.67296253
#>  [6,] 0.45397192 0.45397192 0.45397192
#>  [7,] 0.32237705 0.32237705 0.32237705
#>  [8,] 0.04484445 0.04484445 0.04484445
#>  [9,] 0.29520680 0.15882327 1.15243984
#> [10,] 0.01983528 0.06282899 1.20249288
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04334009 0.04334009 0.04334009
#>  [2,] 0.17561968 0.44658364 0.44658364
#>  [3,] 0.08613742 0.11661793 0.11661793
#>  [4,] 0.03264129 0.03264129 0.03264129
#>  [5,] 0.47090999 0.35734310 0.47090999
#>  [6,] 0.33576693 0.33576693 0.33576693
#>  [7,] 0.70584250 0.70584250 0.70584250
#>  [8,] 0.91949524 0.57804534 0.91949524
#>  [9,] 0.20758698 0.20758698 0.20758698
#> [10,] 0.48328632 0.35833264 0.42537566
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.30331665 0.30331665 0.09585789
#>  [2,] 0.03430155 0.03430155 0.03430155
#>  [3,] 0.57926782 1.11139030 0.60349757
#>  [4,] 0.96035327 0.75224844 0.50244179
#>  [5,] 0.07429128 0.07429128 0.07429128
#>  [6,] 0.16146116 0.16146116 0.16146116
#>  [7,] 1.94611122 0.66581186 0.66581186
#>  [8,] 0.02382250 0.42307353 0.12139863
#>  [9,] 0.70662260 0.70662260 0.70662260
#> [10,] 1.00324739 0.52364711 0.53248135

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.52905595 0.61312631 0.37699333
#>  [2,] 0.35778989 0.45382243 0.49504201
#>  [3,] 0.01055085 0.01055085 0.01055085
#>  [4,] 0.01675871 0.01675871 0.10306900
#>  [5,] 0.40294980 0.23339839 0.40294980
#>  [6,] 0.41077404 0.23723009 0.20603141
#>  [7,] 0.73112666 0.21888868 0.01917225
#>  [8,] 1.15868013 0.04071594 0.57772594
#>  [9,] 0.11687402 0.11687402 0.04624650
#> [10,] 0.66641453 0.25494259 0.25494259
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28339320 1.57436711 1.26775191
#>  [2,] 0.04510718 0.49235840 0.49235840
#>  [3,] 0.17175575 0.28802447 0.31791194
#>  [4,] 0.78548739 0.99153666 0.11455014
#>  [5,] 0.66194766 0.20577952 0.20577952
#>  [6,] 0.39578791 0.14266457 1.03845596
#>  [7,] 0.35843098 0.05217831 0.03407246
#>  [8,] 0.13440243 0.13440243 0.13440243
#>  [9,] 0.29716054 0.14795499 0.14207188
#> [10,] 0.21094693 0.19902927 0.86265554
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24576024 0.54865544 0.17951370
#>  [2,] 0.44891098 0.44891098 0.29384210
#>  [3,] 0.55605629 0.39221041 1.66961113
#>  [4,] 0.96233512 0.79943991 0.08803214
#>  [5,] 0.23715478 0.30579592 0.23715478
#>  [6,] 0.33433085 0.33433085 0.14803250
#>  [7,] 0.68082052 0.68082052 0.06519115
#>  [8,] 0.05957864 0.05957864 0.05957864
#>  [9,] 1.46900486 0.22898646 0.31848343
#> [10,] 0.50776111 0.64291481 0.16143867
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.6326440 1.05880667 1.022793590
#>  [2,] 0.1292728 0.22608680 0.153298761
#>  [3,] 0.3560563 0.13716811 0.112203791
#>  [4,] 0.7201576 0.72015762 1.507224042
#>  [5,] 0.2493767 0.24937671 0.249376709
#>  [6,] 0.1779041 0.89488908 0.008770253
#>  [7,] 0.7177724 0.13789945 0.717772413
#>  [8,] 0.3527229 0.04264436 0.352722936
#>  [9,] 0.2846975 0.28085167 0.816881495
#> [10,] 1.3471269 0.91098668 0.417578032

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.56509805 0.05979506 0.05979506
#>  [2,] 0.69252230 0.07161467 1.82046251
#>  [3,] 0.71551283 0.21373759 0.21373759
#>  [4,] 0.56832096 0.33056075 0.33056075
#>  [5,] 0.28504357 0.28504357 0.12162842
#>  [6,] 0.49626855 0.49626855 0.49626855
#>  [7,] 0.16504782 0.07666654 0.50098719
#>  [8,] 0.08361967 0.01956566 0.08361967
#>  [9,] 0.11332865 0.11332865 0.11332865
#> [10,] 0.19271547 0.12408306 0.59581156
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33435352 0.33435352 0.33435352
#>  [2,] 0.38000635 0.38000635 0.38000635
#>  [3,] 0.24858598 0.24858598 0.24858598
#>  [4,] 0.26544751 0.33821864 0.04737565
#>  [5,] 0.03834268 1.06328937 1.03781050
#>  [6,] 0.76965888 0.46789850 0.79645883
#>  [7,] 0.10524258 0.64494922 0.06975577
#>  [8,] 0.80915312 0.30975886 0.68005914
#>  [9,] 0.48746571 0.48746571 0.02208116
#> [10,] 0.45246133 0.05931007 0.34494480
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.44075976 0.14910045 0.44075976
#>  [2,] 0.46732194 0.56967930 0.46732194
#>  [3,] 0.12640187 0.12640187 0.12640187
#>  [4,] 0.01927509 0.59347394 0.06553024
#>  [5,] 0.05043990 0.05043990 0.05043990
#>  [6,] 0.16509604 0.16509604 0.16509604
#>  [7,] 0.19730519 0.77752556 0.17582203
#>  [8,] 0.52800211 0.93458377 0.11893655
#>  [9,] 0.23446707 1.07611116 1.07611116
#> [10,] 0.31160263 0.06953017 0.14759227
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.50489513 0.03683558 1.06717889
#>  [2,] 0.48015506 0.33211989 0.48015506
#>  [3,] 0.15608834 0.15608834 0.15608834
#>  [4,] 0.19307226 0.19307226 0.19307226
#>  [5,] 0.90420090 0.90420090 0.64254352
#>  [6,] 1.06564359 0.03272523 0.07935944
#>  [7,] 0.15675425 0.39661825 0.88258543
#>  [8,] 0.11872463 0.11872463 0.11872463
#>  [9,] 0.02682207 0.79058471 0.03957540
#> [10,] 0.51520281 0.22604510 0.51520281

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.002369928 0.002369928 0.002369928
#>  [2,] 0.066841312 1.355933794 0.040627826
#>  [3,] 0.080678283 0.184482026 0.033223143
#>  [4,] 0.088308733 0.088308733 0.088308733
#>  [5,] 0.028759275 0.328537886 0.371048831
#>  [6,] 0.171491189 0.228844756 0.654330571
#>  [7,] 0.616292480 1.249103399 0.458289728
#>  [8,] 0.546628772 0.373349210 0.373349210
#>  [9,] 0.106969629 0.106969629 0.106969629
#> [10,] 0.255200021 0.557278556 0.239224137
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.70876134 0.018734604 0.01873460
#>  [2,] 0.61518378 0.579608039 0.02556226
#>  [3,] 0.08909346 0.007885001 0.02407270
#>  [4,] 0.11577166 0.252383386 0.25238339
#>  [5,] 0.04890323 0.048903227 0.23140870
#>  [6,] 0.40977171 0.022137031 0.02213703
#>  [7,] 0.10132992 0.136398415 0.76211427
#>  [8,] 0.58582916 0.585829156 0.45359902
#>  [9,] 0.20380632 0.203806318 0.20380632
#> [10,] 0.25621447 0.033470521 0.09294874
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.25958937 0.25958937 0.14100602
#>  [2,] 0.54478066 0.17343657 0.17343657
#>  [3,] 0.06123966 0.06123966 0.06123966
#>  [4,] 0.08626997 0.03928855 0.19252601
#>  [5,] 0.52587944 0.33340775 0.33340775
#>  [6,] 0.10764697 0.19930721 0.19930721
#>  [7,] 0.57736756 0.57736756 0.18925942
#>  [8,] 0.05381375 0.05381375 0.05381375
#>  [9,] 0.39410283 0.91217644 0.50911393
#> [10,] 0.17911110 0.17911110 0.17911110
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>               [,1]       [,2]        [,3]
#>  [1,] 0.0804582437 0.17510695 0.080458244
#>  [2,] 0.2307656282 0.19955877 0.199558767
#>  [3,] 0.6712355796 0.02778706 0.050140500
#>  [4,] 0.0001699461 0.07610244 0.364586220
#>  [5,] 0.0096890393 0.54180398 0.009689039
#>  [6,] 0.2531147788 0.57359796 0.967425168
#>  [7,] 0.3949385502 0.16945335 0.443610225
#>  [8,] 0.2308016655 0.15734834 0.230801666
#>  [9,] 0.3440332714 0.34403327 0.344033271
#> [10,] 0.1123743975 0.97856912 0.978569125