Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.94543913 0.4588255 3.20299909
#>  [2,] 2.25973367 2.4941084 0.01176162
#>  [3,] 0.23036348 1.0916376 0.69274714
#>  [4,] 0.45569029 2.1311802 1.72315975
#>  [5,] 3.55898712 3.3479713 5.53318943
#>  [6,] 0.56716349 0.5671635 0.56716349
#>  [7,] 3.23985039 3.2398504 0.60961063
#>  [8,] 0.60720976 0.6072098 0.60720976
#>  [9,] 0.83017206 0.8301721 0.67058040
#> [10,] 0.07928028 3.0563311 1.27363733
rpextmo(10, 3, a = 1)      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.730769174 0.730769174 0.730769174
#>  [2,] 0.640944998 0.640944998 0.640944998
#>  [3,] 1.010968487 1.010968487 1.010968487
#>  [4,] 0.935004157 0.935004157 0.935004157
#>  [5,] 0.638391863 0.638391863 0.638391863
#>  [6,] 0.004630708 0.004630708 0.004630708
#>  [7,] 0.126671677 0.126671677 0.126671677
#>  [8,] 0.732606845 0.732606845 0.732606845
#>  [9,] 3.777577525 3.777577525 3.777577525
#> [10,] 0.419798830 0.419798830 0.419798830
rpextmo(10, 3, b = 1)      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1491606 1.01386901 1.16710649
#>  [2,] 1.3819405 1.22729171 0.20020350
#>  [3,] 0.2660340 0.18062049 0.72589503
#>  [4,] 0.1802058 1.09170451 1.02544645
#>  [5,] 0.6113865 0.88590804 1.80313351
#>  [6,] 0.3438946 0.07140455 0.76556727
#>  [7,] 0.6022739 0.31828484 0.91819764
#>  [8,] 0.2017364 3.75821716 2.60648892
#>  [9,] 0.6490831 1.32884042 1.60562089
#> [10,] 0.6899357 0.74347524 0.01101318

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.17788123 0.3772963 0.3772963
#>  [2,] 0.01425517 1.0171374 1.3323577
#>  [3,] 1.02159849 1.0215985 1.0215985
#>  [4,] 0.89938393 0.1077749 2.4642184
#>  [5,] 1.97303974 1.9730397 1.4650261
#>  [6,] 1.42735030 4.7754127 0.1953367
#>  [7,] 1.22795591 1.2279559 1.2279559
#>  [8,] 1.60985657 1.0996194 0.3664782
#>  [9,] 2.32959574 3.2083174 0.5801476
#> [10,] 1.84461228 2.7185773 0.3795365
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.10308510 0.10308510 0.10308510
#>  [2,] 0.77546221 0.77546221 0.77546221
#>  [3,] 0.42151900 0.42151900 0.42151900
#>  [4,] 1.00098219 1.00098219 1.00098219
#>  [5,] 1.79830441 1.79830441 1.79830441
#>  [6,] 0.54425924 0.54425924 0.54425924
#>  [7,] 0.84653585 0.84653585 0.84653585
#>  [8,] 0.09500901 0.09500901 0.09500901
#>  [9,] 0.37949256 0.37949256 0.37949256
#> [10,] 0.37095311 0.37095311 0.37095311
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4690844 0.6704716 0.5266065
#>  [2,] 2.2522728 0.4782704 0.6365212
#>  [3,] 2.6449112 0.4571895 2.6225053
#>  [4,] 0.2350589 0.4807460 0.1316633
#>  [5,] 0.3659450 1.3994800 0.3586521
#>  [6,] 0.3182971 0.5191232 2.2477545
#>  [7,] 1.9857700 0.4547351 1.3711246
#>  [8,] 0.7719829 0.9097203 0.9309004
#>  [9,] 1.2414690 0.9991790 0.7963436
#> [10,] 0.9983534 2.5232684 1.4049165

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.15349609 0.6141810 1.6378532
#>  [2,] 0.93752736 2.8207493 1.8282928
#>  [3,] 0.26031008 0.2603101 0.2603101
#>  [4,] 0.15476603 0.1547660 0.1547660
#>  [5,] 0.75045208 0.7504521 0.7504521
#>  [6,] 0.07996221 1.3192284 0.8236983
#>  [7,] 2.91142362 0.2122239 2.9361907
#>  [8,] 2.72478991 2.4542173 0.5135483
#>  [9,] 0.58581813 0.5858181 0.1195432
#> [10,] 1.11952355 0.9507268 3.3834382
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7399913 0.7399913 0.7399913
#>  [2,] 0.1913870 0.1913870 0.1913870
#>  [3,] 0.9123118 0.9123118 0.9123118
#>  [4,] 1.3817818 1.3817818 1.3817818
#>  [5,] 0.8267904 0.8267904 0.8267904
#>  [6,] 1.3780968 1.3780968 1.3780968
#>  [7,] 0.1098137 0.1098137 0.1098137
#>  [8,] 3.3233283 3.3233283 3.3233283
#>  [9,] 1.0490924 1.0490924 1.0490924
#> [10,] 0.2436139 0.2436139 0.2436139
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7373749 0.2552306 0.6674670
#>  [2,] 0.1205867 0.2855661 0.4013303
#>  [3,] 4.9183263 0.3217939 2.7661687
#>  [4,] 0.7764162 1.9371230 1.2151278
#>  [5,] 3.0450654 0.9664959 2.4180963
#>  [6,] 1.2973323 2.7052612 0.3166365
#>  [7,] 3.2221034 0.5291990 0.3809846
#>  [8,] 0.1286712 0.7054863 0.0322473
#>  [9,] 0.1499472 0.4061420 0.5851500
#> [10,] 2.0274672 1.7481817 1.8727913

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9303130 0.2640668 0.37430557
#>  [2,] 1.0435457 0.9697019 1.24021360
#>  [3,] 1.7418840 0.5126276 2.17061106
#>  [4,] 0.7011587 1.9421376 1.90491744
#>  [5,] 1.7166384 1.2873206 0.70915893
#>  [6,] 1.4025514 0.2548839 0.03331646
#>  [7,] 0.1048074 0.1048074 0.10480738
#>  [8,] 2.1289266 6.5528327 1.58835871
#>  [9,] 1.8382993 1.7981173 0.23572963
#> [10,] 0.6310194 0.6310194 0.63101943
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12780512 0.12780512 0.12780512
#>  [2,] 0.57356708 0.57356708 0.57356708
#>  [3,] 1.27323941 1.27323941 1.27323941
#>  [4,] 0.87088027 0.87088027 0.87088027
#>  [5,] 0.86829537 0.86829537 0.86829537
#>  [6,] 0.20772138 0.20772138 0.20772138
#>  [7,] 1.22007858 1.22007858 1.22007858
#>  [8,] 0.03538975 0.03538975 0.03538975
#>  [9,] 1.64317161 1.64317161 1.64317161
#> [10,] 0.31067600 0.31067600 0.31067600
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.60549485 1.27582438 0.5091716
#>  [2,] 0.03666172 1.20160707 1.1770737
#>  [3,] 0.42536469 0.84716847 1.2092830
#>  [4,] 3.98254994 0.06377547 0.3473127
#>  [5,] 2.79688237 0.91780488 0.3679324
#>  [6,] 1.04666358 0.41750920 0.9496642
#>  [7,] 1.05003306 0.62712814 1.8361698
#>  [8,] 2.06651775 2.76052864 0.8876074
#>  [9,] 2.31456974 0.64641792 0.9498489
#> [10,] 0.09916405 0.72472540 0.1855615

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,]  1.9696900 0.77037673 1.0303301
#>  [2,]  0.3883340 4.12943470 0.2110415
#>  [3,]  0.5514930 1.68242348 1.4848630
#>  [4,] 15.1435172 0.85369175 5.3327213
#>  [5,]  2.0818612 0.11339138 5.8656526
#>  [6,]  0.6373410 0.81779974 0.1839059
#>  [7,]  0.5996511 0.09153130 0.5996511
#>  [8,]  0.1659987 1.04670382 1.0467038
#>  [9,]  0.7917265 1.53829444 1.5382944
#> [10,]  0.2825298 0.02572943 1.5265485
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2567741 0.2567741 0.2567741
#>  [2,] 0.7602613 0.7602613 0.7602613
#>  [3,] 0.2975723 0.2975723 0.2975723
#>  [4,] 0.4275582 0.4275582 0.4275582
#>  [5,] 0.1595468 0.1595468 0.1595468
#>  [6,] 0.5658345 0.5658345 0.5658345
#>  [7,] 0.4402609 0.4402609 0.4402609
#>  [8,] 0.7953836 0.7953836 0.7953836
#>  [9,] 0.5209042 0.5209042 0.5209042
#> [10,] 0.2343271 0.2343271 0.2343271
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8660471 0.19040595 0.1356484
#>  [2,] 1.8750306 0.42155542 0.5729494
#>  [3,] 0.2660883 0.61393064 0.4719412
#>  [4,] 0.5365813 0.53782225 0.1866779
#>  [5,] 0.6532920 0.11444856 3.5075200
#>  [6,] 0.5380041 0.04051520 0.5853413
#>  [7,] 0.4322544 0.84967432 1.2235973
#>  [8,] 0.7388517 0.08282599 0.1021965
#>  [9,] 0.8859026 2.14822439 0.7222590
#> [10,] 1.5879220 0.13838466 6.1651133

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7211383 2.4918740 4.9489801
#>  [2,] 0.4661267 0.4661267 0.4661267
#>  [3,] 3.4461257 4.4457261 6.4026500
#>  [4,] 1.8956941 1.8956941 0.4142092
#>  [5,] 2.5844314 0.4771175 0.4323677
#>  [6,] 6.5475097 1.5341265 0.2315582
#>  [7,] 1.5308800 1.1216197 1.3266695
#>  [8,] 2.3275150 0.5299412 2.2518466
#>  [9,] 2.0966133 0.9374481 0.3998076
#> [10,] 0.4166149 0.4166149 0.4166149
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4065272 0.4065272 0.4065272
#>  [2,] 0.2648138 0.2648138 0.2648138
#>  [3,] 0.3087854 0.3087854 0.3087854
#>  [4,] 0.5268449 0.5268449 0.5268449
#>  [5,] 1.3198531 1.3198531 1.3198531
#>  [6,] 0.6900042 0.6900042 0.6900042
#>  [7,] 0.2114570 0.2114570 0.2114570
#>  [8,] 0.9315004 0.9315004 0.9315004
#>  [9,] 0.1291786 0.1291786 0.1291786
#> [10,] 0.9555728 0.9555728 0.9555728
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4429706 1.9702717 0.5143306
#>  [2,] 1.1043384 0.3989870 0.4498979
#>  [3,] 0.6021424 0.1806461 0.5657524
#>  [4,] 0.2423718 0.6432984 1.4217287
#>  [5,] 1.0787177 0.3016285 1.5847590
#>  [6,] 2.0672060 2.0974571 1.7033883
#>  [7,] 0.4485488 0.8456504 0.2606761
#>  [8,] 0.7596875 0.4454329 1.5785268
#>  [9,] 2.3818333 0.7383638 0.1542763
#> [10,] 0.7148016 2.1885005 0.2920517

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9857302 3.0999489 0.6545245
#>  [2,] 0.9742080 1.1523168 0.4232038
#>  [3,] 0.7846132 1.1801341 0.0535128
#>  [4,] 2.2880245 1.5637564 1.9930759
#>  [5,] 0.5407279 1.9040906 0.8660880
#>  [6,] 0.5663542 0.9278652 0.8713691
#>  [7,] 0.5592339 1.5129164 0.7684468
#>  [8,] 1.6914387 1.0450694 0.9005151
#>  [9,] 0.8523701 0.4114266 0.4579431
#> [10,] 0.1866575 0.1866575 0.1866575
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.0486680 0.0486680 0.0486680
#>  [2,] 0.3181771 0.3181771 0.3181771
#>  [3,] 1.2070226 1.2070226 1.2070226
#>  [4,] 0.1203516 0.1203516 0.1203516
#>  [5,] 2.4424560 2.4424560 2.4424560
#>  [6,] 0.2401410 0.2401410 0.2401410
#>  [7,] 2.0983453 2.0983453 2.0983453
#>  [8,] 0.6057813 0.6057813 0.6057813
#>  [9,] 0.4040354 0.4040354 0.4040354
#> [10,] 1.0810700 1.0810700 1.0810700
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2394991 1.7376760 0.05794656
#>  [2,] 0.3512726 1.7376207 1.43730802
#>  [3,] 0.9500816 0.9354046 1.25746075
#>  [4,] 0.4927699 1.6502388 0.19636882
#>  [5,] 0.8953083 2.1680772 0.32513121
#>  [6,] 1.0096409 0.7596275 3.78682045
#>  [7,] 1.0691185 1.7486489 0.40182677
#>  [8,] 0.3877325 2.1906951 0.28107958
#>  [9,] 1.2746641 0.7239870 0.81553157
#> [10,] 0.2093051 0.2373365 3.04832333

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 2.2359776 1.62173242 2.6492382
#>  [2,] 3.2232851 0.01530055 4.0719333
#>  [3,] 1.7865531 1.32110734 0.4206643
#>  [4,] 0.8898992 5.22290783 0.3789787
#>  [5,] 3.2040952 1.68728398 1.8402447
#>  [6,] 4.5183261 2.40368112 2.4638006
#>  [7,] 1.0743941 1.07439408 1.0743941
#>  [8,] 1.0142165 0.32683296 1.8566106
#>  [9,] 0.3267934 0.30823635 0.4194964
#> [10,] 0.3909813 0.61651801 1.3430253
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07229045 0.07229045 0.07229045
#>  [2,] 1.74742244 1.74742244 1.74742244
#>  [3,] 0.65497653 0.65497653 0.65497653
#>  [4,] 0.12803462 0.12803462 0.12803462
#>  [5,] 0.23332760 0.23332760 0.23332760
#>  [6,] 0.78816839 0.78816839 0.78816839
#>  [7,] 4.18439983 4.18439983 4.18439983
#>  [8,] 0.13764305 0.13764305 0.13764305
#>  [9,] 0.66385113 0.66385113 0.66385113
#> [10,] 0.10389304 0.10389304 0.10389304
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>              [,1]       [,2]       [,3]
#>  [1,] 0.369871478 1.42999198 0.64737164
#>  [2,] 1.732540483 0.37765001 0.51026840
#>  [3,] 0.132071850 0.46929472 0.17668205
#>  [4,] 1.687015813 0.33217620 0.02141388
#>  [5,] 0.865174850 0.04897898 1.13417035
#>  [6,] 0.372808667 0.07648310 0.65965475
#>  [7,] 0.764768548 0.48199335 1.13706614
#>  [8,] 0.682621575 1.49681942 2.11746366
#>  [9,] 0.608355719 0.15005078 1.14417458
#> [10,] 0.009201909 0.68764655 0.04060784

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.12003327 1.0256437 0.32772401
#>  [2,] 3.06868740 0.9370903 1.42047021
#>  [3,] 0.16409963 0.2397111 0.23971107
#>  [4,] 4.68393381 1.8543349 2.18249742
#>  [5,] 0.37257992 1.2578276 0.08254236
#>  [6,] 0.06850345 0.2315047 0.23150474
#>  [7,] 0.78108166 2.4033017 2.14956914
#>  [8,] 3.18607631 0.3981231 3.18607631
#>  [9,] 4.48973795 0.7656141 6.37007953
#> [10,] 3.00030802 3.5122989 2.34600159
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32694388 0.32694388 0.32694388
#>  [2,] 0.68899379 0.68899379 0.68899379
#>  [3,] 0.74748628 0.74748628 0.74748628
#>  [4,] 0.75438867 0.75438867 0.75438867
#>  [5,] 0.92628474 0.92628474 0.92628474
#>  [6,] 0.13553861 0.13553861 0.13553861
#>  [7,] 0.32818654 0.32818654 0.32818654
#>  [8,] 0.27886316 0.27886316 0.27886316
#>  [9,] 0.42009803 0.42009803 0.42009803
#> [10,] 0.09239814 0.09239814 0.09239814
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>             [,1]       [,2]        [,3]
#>  [1,] 4.21521529 0.06420191 0.196163910
#>  [2,] 0.24679332 0.96452156 0.138359409
#>  [3,] 0.91659334 0.13961769 0.373044671
#>  [4,] 0.89078071 0.29425279 1.630713196
#>  [5,] 0.65548577 2.06849716 0.151708030
#>  [6,] 0.01968719 0.19730110 2.555379838
#>  [7,] 0.03530120 1.05202988 0.571167582
#>  [8,] 0.23310205 0.19154376 3.936067205
#>  [9,] 0.32052538 0.37290741 0.537768936
#> [10,] 0.53105523 0.82543659 0.004724474

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>               [,1]         [,2]         [,3]
#>  [1,] 3.4210277538 4.6471353730 0.8001738929
#>  [2,] 2.9975537527 3.6025290918 0.9033551582
#>  [3,] 1.7420111315 0.3131863138 1.4084891713
#>  [4,] 0.8783921212 0.3495127268 0.1231431167
#>  [5,] 0.0693087411 0.7510149807 1.2531343973
#>  [6,] 0.2064655858 1.9020356258 1.8704119456
#>  [7,] 0.4517885259 0.3133174528 1.3673983279
#>  [8,] 0.2548033059 0.2548033059 0.0361870659
#>  [9,] 1.2195384918 1.3650853800 1.3650853800
#> [10,] 0.0003003587 0.0003003587 0.0003003587
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20445351 0.20445351 0.20445351
#>  [2,] 0.02755600 0.02755600 0.02755600
#>  [3,] 0.43307520 0.43307520 0.43307520
#>  [4,] 1.26279941 1.26279941 1.26279941
#>  [5,] 3.06560322 3.06560322 3.06560322
#>  [6,] 0.68184350 0.68184350 0.68184350
#>  [7,] 0.30135953 0.30135953 0.30135953
#>  [8,] 0.19179223 0.19179223 0.19179223
#>  [9,] 0.28153605 0.28153605 0.28153605
#> [10,] 0.08567939 0.08567939 0.08567939
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 2.2843675 4.08440779 0.19667340
#>  [2,] 1.2098697 0.01176178 2.59992726
#>  [3,] 0.8965696 0.09616366 1.03382408
#>  [4,] 0.1028163 4.23940427 0.09971357
#>  [5,] 1.0780415 0.20200109 1.93004355
#>  [6,] 0.5990514 0.22260846 0.81337924
#>  [7,] 3.0152277 0.01001732 0.21488454
#>  [8,] 1.0308235 0.78726336 0.42417732
#>  [9,] 1.0045494 0.82214214 3.20661185
#> [10,] 1.4644995 0.12888561 1.24108606

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17271268 0.17271268 0.17271268
#>  [2,] 1.66800490 0.06555463 1.00443330
#>  [3,] 0.77045133 0.47996287 0.47996287
#>  [4,] 0.09902565 0.02194841 0.17619526
#>  [5,] 0.32737588 1.17276027 1.17276027
#>  [6,] 0.12246768 1.11458668 0.02203234
#>  [7,] 0.06997699 0.06997699 0.06997699
#>  [8,] 0.87082395 1.27688017 0.72488484
#>  [9,] 0.10914438 0.32360162 0.41439770
#> [10,] 0.35814118 0.11008745 0.27616013
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37637868 0.81050634 0.11921577
#>  [2,] 0.12927229 1.08454226 1.66104574
#>  [3,] 0.74588563 0.74588563 2.60443074
#>  [4,] 0.12890632 0.12890632 0.05958639
#>  [5,] 0.01323659 0.01323659 0.01323659
#>  [6,] 0.18820526 0.18820526 0.14395487
#>  [7,] 0.21671511 1.51002438 2.26263069
#>  [8,] 0.03474245 0.10804228 0.65916937
#>  [9,] 0.44599340 0.09178398 2.59110128
#> [10,] 0.64171453 0.49073988 5.16249359
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.622431998 0.1383602 0.1383602
#>  [2,] 0.385495700 0.7441745 0.1989968
#>  [3,] 0.016773542 0.8247357 0.9847240
#>  [4,] 0.004053543 0.4142536 0.2304735
#>  [5,] 0.650015490 0.2375626 0.4165842
#>  [6,] 0.252903202 1.0095241 0.2529032
#>  [7,] 0.019209452 2.2523948 0.4466878
#>  [8,] 0.283366286 0.2833663 0.1363414
#>  [9,] 0.964398201 1.4616689 0.3169570
#> [10,] 1.912904098 0.9751600 0.2709896
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4280071 0.42800713 2.83502461
#>  [2,] 0.8911536 0.87092185 0.45204367
#>  [3,] 0.2938663 0.29671422 0.87234740
#>  [4,] 0.2557348 0.75092248 0.25573476
#>  [5,] 0.2240469 0.06716223 0.22404691
#>  [6,] 0.2261719 0.22617187 1.17576650
#>  [7,] 0.1628422 0.52328003 0.24841183
#>  [8,] 0.1883756 0.18837556 0.04145731
#>  [9,] 0.7726774 0.47395749 0.77267743
#> [10,] 0.7995936 0.40804195 0.25782807
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.13175959 0.3810273 0.19548061
#>  [2,] 0.04556637 0.4801896 1.04956663
#>  [3,] 0.59879957 1.1105780 1.45151895
#>  [4,] 0.47943941 1.2400223 0.49264145
#>  [5,] 0.60937132 1.4422965 1.39475814
#>  [6,] 0.46790783 1.4214064 0.01959345
#>  [7,] 0.95693907 0.7831760 1.93947794
#>  [8,] 3.64494786 3.0753577 3.64494786
#>  [9,] 1.58024934 0.5189846 0.34615481
#> [10,] 0.72981364 0.4436035 0.86594221

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6959870 0.4792737 2.45787512
#>  [2,] 0.8768598 1.4221819 0.08967903
#>  [3,] 0.3466707 0.8994178 1.51120833
#>  [4,] 0.7746358 2.1299800 6.36074409
#>  [5,] 0.2251018 0.2251018 0.22510181
#>  [6,] 3.0825273 1.6013582 2.77210897
#>  [7,] 0.2336264 1.1297661 0.74378025
#>  [8,] 1.0155297 1.0155297 1.01552966
#>  [9,] 1.7407771 1.3014067 3.70746580
#> [10,] 0.8061356 0.7894867 0.80613558
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.66804225 0.6680422 0.6033934
#>  [2,] 0.01746359 2.2702461 2.9813910
#>  [3,] 1.10121211 5.8488738 1.0741174
#>  [4,] 0.94916292 0.0539740 0.1358011
#>  [5,] 2.99603472 0.1289805 5.6501513
#>  [6,] 0.95238572 2.4409470 0.7237819
#>  [7,] 0.25758947 0.2575895 0.2575895
#>  [8,] 0.18578194 0.1857819 0.1857819
#>  [9,] 1.04114765 2.1787852 0.4357765
#> [10,] 1.22268601 2.2837484 2.2837484
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5594538 2.4040192 0.1290252
#>  [2,] 1.1814170 2.0577160 2.1749717
#>  [3,] 1.2095711 0.9872466 1.0076712
#>  [4,] 1.5696364 1.6713086 0.1684265
#>  [5,] 1.2391389 1.2391389 1.2391389
#>  [6,] 0.3678738 1.1420243 0.3496645
#>  [7,] 3.5938241 3.5938241 0.6717490
#>  [8,] 1.1292531 2.8866762 0.6366784
#>  [9,] 0.2817032 0.2817032 0.2817032
#> [10,] 0.0609617 0.0609617 0.0609617
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4217538 0.2171074 1.01690267
#>  [2,] 1.0446540 1.0446540 1.04465398
#>  [3,] 0.4643788 0.5168498 0.17244934
#>  [4,] 1.0898727 0.6520817 2.13309308
#>  [5,] 2.9451964 1.3742218 1.23250921
#>  [6,] 1.8095675 1.8095675 1.80956752
#>  [7,] 0.2899485 0.5606166 0.56061659
#>  [8,] 0.5892028 0.9021346 0.01555785
#>  [9,] 0.8170514 0.2209693 1.15984137
#> [10,] 2.2726308 0.5417883 0.34356993
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.5450394 2.74915808 0.37574644
#>  [2,] 1.1356860 1.92980868 0.24375585
#>  [3,] 0.1371721 0.21288560 1.19077146
#>  [4,] 3.2709215 1.25175939 2.86965114
#>  [5,] 3.8671194 3.86711944 3.38083309
#>  [6,] 0.3141948 2.91913461 5.56349224
#>  [7,] 0.8713195 0.25259588 0.87131945
#>  [8,] 1.6764827 1.83228721 9.70318847
#>  [9,] 2.7770729 0.10962838 0.79944882
#> [10,] 3.2788461 0.09509641 0.04993019

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28395890 0.28395890 0.28395890
#>  [2,] 1.28157601 0.38686911 0.38686911
#>  [3,] 0.88080694 0.39018147 0.88080694
#>  [4,] 0.01960700 0.01960700 0.01960700
#>  [5,] 0.02434829 0.02434829 0.02434829
#>  [6,] 0.51474480 0.51474480 0.27143297
#>  [7,] 0.64139324 1.02059852 0.64139324
#>  [8,] 0.92329588 0.92329588 0.92329588
#>  [9,] 0.09863194 0.09863194 0.09863194
#> [10,] 0.46125797 0.46125797 0.54266609
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.276210145 0.276210145 0.276210145
#>  [2,] 0.457446064 0.457446064 0.457446064
#>  [3,] 0.053407300 0.053407300 0.053407300
#>  [4,] 0.004636433 0.004636433 0.004636433
#>  [5,] 0.061728040 0.304318877 0.061728040
#>  [6,] 0.899275501 0.851596109 0.572756520
#>  [7,] 0.628788536 0.186394646 0.215860525
#>  [8,] 1.115806581 1.298292677 0.563248174
#>  [9,] 0.441385130 0.441385130 0.441385130
#> [10,] 0.027341356 0.413685881 0.516998652
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.71679642 0.71679642 0.7167964
#>  [2,] 0.12043524 0.12043524 0.1204352
#>  [3,] 0.45330134 0.09984465 0.4533013
#>  [4,] 0.15313360 0.15313360 0.1531336
#>  [5,] 0.70288186 0.70288186 0.7028819
#>  [6,] 0.65337913 0.79441109 0.6533791
#>  [7,] 0.67009485 0.18482702 0.4118948
#>  [8,] 0.54016192 0.54016192 0.5401619
#>  [9,] 0.40746235 2.16547363 0.8302713
#> [10,] 0.05151685 0.05151685 0.1544590
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.8419062 0.24297440 0.28804102
#>  [2,] 0.5663191 1.07446889 0.60499937
#>  [3,] 0.5286646 0.52866459 0.52866459
#>  [4,] 0.8523750 0.85237495 1.13006533
#>  [5,] 0.6692745 0.01157605 0.01157605
#>  [6,] 0.4711025 0.02706773 0.02706773
#>  [7,] 0.8830412 0.88304123 0.24856624
#>  [8,] 0.7243075 0.72430755 0.06599301
#>  [9,] 0.3791934 0.37919338 0.37919338
#> [10,] 0.3666138 0.05489227 0.05489227
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.59049258 2.1158501 0.5904926
#>  [2,] 0.12732280 0.1273228 0.1273228
#>  [3,] 0.48905205 0.4890521 0.4890521
#>  [4,] 0.09008651 0.2715861 0.5225688
#>  [5,] 0.24680066 0.9350573 0.9350573
#>  [6,] 0.15639444 0.8993687 0.1563944
#>  [7,] 0.05449311 0.0687426 0.0687426
#>  [8,] 0.20622754 0.2062275 0.2062275
#>  [9,] 0.17623263 0.1762326 0.1762326
#> [10,] 0.80686094 0.2483651 0.2118772

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07634762 0.96406341 0.76209858
#>  [2,] 0.42394525 0.22757789 1.15953161
#>  [3,] 0.02570286 0.28117713 0.09381153
#>  [4,] 0.22294429 0.13113945 0.13113945
#>  [5,] 0.03418307 0.07695749 0.92433763
#>  [6,] 0.81032365 0.81032365 0.81032365
#>  [7,] 0.12935170 0.12935170 0.12935170
#>  [8,] 0.01892997 0.27504689 0.25307936
#>  [9,] 0.29140852 0.29140852 0.29140852
#> [10,] 0.09984213 0.04347140 0.75789614
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4340860 0.43408601 0.8859496
#>  [2,] 0.9793775 1.05376266 0.5614114
#>  [3,] 0.1018784 0.04693947 0.1018784
#>  [4,] 0.1565833 0.15658326 0.1565833
#>  [5,] 1.6172979 0.39620309 1.4577642
#>  [6,] 0.2317618 0.40886907 0.7384041
#>  [7,] 0.2795624 0.27956242 0.2795624
#>  [8,] 0.1872335 0.33960089 0.3396009
#>  [9,] 0.5572435 0.58737930 0.6361861
#> [10,] 1.5464282 0.50743741 1.3133893
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.005578839 0.84585438 0.84585438
#>  [2,] 0.251792950 0.23626923 0.52711265
#>  [3,] 0.351285029 0.67808783 0.93172096
#>  [4,] 0.011296082 0.01129608 0.01129608
#>  [5,] 0.095325206 0.08984352 0.00704197
#>  [6,] 0.665242698 0.50815752 0.66524270
#>  [7,] 0.674623712 0.26322957 0.28605256
#>  [8,] 1.339703241 1.10110258 1.36294958
#>  [9,] 0.064856918 0.89174903 0.32159159
#> [10,] 0.259270992 0.91688205 0.44757846
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>               [,1]         [,2]         [,3]
#>  [1,] 0.3883491726 0.1690792646 0.2133559506
#>  [2,] 0.0149937467 0.1154258107 0.1154258107
#>  [3,] 0.4374706187 0.4374706187 0.4374706187
#>  [4,] 0.0967498080 0.3735172723 0.3735172723
#>  [5,] 0.0493779619 0.5227996657 0.1868997128
#>  [6,] 0.4788171930 0.0462781607 0.3032260504
#>  [7,] 1.0313432041 0.2303096099 0.0433854299
#>  [8,] 0.2478161119 0.0546782335 0.2478161119
#>  [9,] 0.2772565915 0.6204598693 0.6204598693
#> [10,] 0.0006591401 0.0006591401 0.0006591401

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.13199706 0.18217736 0.1821774
#>  [2,] 0.41279026 0.85854852 0.1604622
#>  [3,] 0.02385794 0.90748381 0.6689188
#>  [4,] 0.89409132 0.06368445 0.7081366
#>  [5,] 0.97071238 0.54795364 0.1425697
#>  [6,] 0.32037866 0.32037866 0.1490566
#>  [7,] 0.46322281 0.46322281 0.1821783
#>  [8,] 0.26818526 0.26818526 0.2681853
#>  [9,] 0.43736814 0.73186545 0.7318654
#> [10,] 0.06087943 0.06087943 0.2102315
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.047818216 0.04781822 0.04781822
#>  [2,] 0.197327341 0.19732734 0.19732734
#>  [3,] 1.210891478 0.83744105 0.68752678
#>  [4,] 1.109204770 1.06278301 0.42269501
#>  [5,] 0.745197184 0.69582456 0.78531295
#>  [6,] 0.005021027 0.99643446 0.81266836
#>  [7,] 0.218855902 0.21885590 0.21885590
#>  [8,] 0.605853293 0.19210255 0.29934291
#>  [9,] 0.366396002 0.33357102 0.58960588
#> [10,] 0.016529426 0.01652943 0.24217824
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04176326 0.04176326 0.04176326
#>  [2,] 0.34463559 0.34463559 0.34463559
#>  [3,] 0.28199014 1.15953252 0.69201959
#>  [4,] 0.54617605 0.54617605 0.02238323
#>  [5,] 0.35342354 0.35342354 0.35342354
#>  [6,] 0.06699417 0.17716549 0.35435365
#>  [7,] 0.14351377 0.02786880 0.02786880
#>  [8,] 0.55697107 2.17661247 0.00864654
#>  [9,] 1.19725754 2.12753382 0.37951611
#> [10,] 0.11606451 0.11606451 0.11606451
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26854718 0.26854718 0.33495773
#>  [2,] 0.08776916 0.08776916 0.05965965
#>  [3,] 0.65273669 0.26767076 0.65273669
#>  [4,] 0.63707969 0.06970390 0.40240084
#>  [5,] 0.20790039 0.20790039 0.20790039
#>  [6,] 0.58822983 0.01893180 0.65032506
#>  [7,] 0.04378930 0.04378930 0.04378930
#>  [8,] 0.25347704 0.15494478 0.56499384
#>  [9,] 0.14220595 0.03701270 0.36631278
#> [10,] 0.11777639 0.09514983 0.11777639

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21298926 0.01697485 0.19146991
#>  [2,] 0.72725947 0.13729590 0.72917929
#>  [3,] 0.11441990 0.23201808 0.23201808
#>  [4,] 0.15142143 0.49629495 0.04427048
#>  [5,] 0.04377800 0.08773228 0.04013328
#>  [6,] 0.54270684 1.16170563 0.90327688
#>  [7,] 0.22958153 0.22958153 0.22958153
#>  [8,] 0.06144964 0.06144964 0.06144964
#>  [9,] 0.08544437 0.01826920 0.01826920
#> [10,] 0.42425964 0.43402532 0.43402532
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05664881 0.37523545 0.96213162
#>  [2,] 0.07310106 0.19313721 0.19313721
#>  [3,] 0.31639048 0.02252396 0.02252396
#>  [4,] 0.78857610 0.31777829 0.28223124
#>  [5,] 0.32583690 0.06196534 0.00872108
#>  [6,] 0.03352113 1.04653951 0.84892375
#>  [7,] 1.24349511 0.64903643 0.31528641
#>  [8,] 0.24638650 0.24638650 0.24638650
#>  [9,] 0.09021567 0.02418726 0.08110038
#> [10,] 0.07265223 0.06270403 0.08431564
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.48054368 0.48054368 0.48054368
#>  [2,] 0.48594446 0.84254647 0.01053277
#>  [3,] 0.67426270 0.02602037 0.67261274
#>  [4,] 0.04987608 0.27859374 0.65570214
#>  [5,] 0.12476281 0.12548151 0.12548151
#>  [6,] 0.05000089 0.05000089 0.05000089
#>  [7,] 0.37214445 0.37214445 0.37214445
#>  [8,] 0.54159499 0.54159499 1.49110267
#>  [9,] 0.49003385 0.30382227 0.42684837
#> [10,] 0.29266432 0.28420103 0.28298864
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.45011898 0.56235544 0.56235544
#>  [2,] 1.15282775 1.96905905 2.40195833
#>  [3,] 0.70872806 0.70872806 0.14957290
#>  [4,] 0.03192410 0.53391066 0.38633839
#>  [5,] 0.10763819 0.10763819 0.10763819
#>  [6,] 0.04145707 0.11178316 0.11178316
#>  [7,] 0.19876040 0.10560882 0.70437694
#>  [8,] 0.51978171 0.08372295 0.15450458
#>  [9,] 0.36569428 0.36569428 0.20554420
#> [10,] 0.10765931 0.10765931 0.04330335