Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.59206319 0.5920632 0.1406121
#>  [2,] 4.50904689 1.4304133 0.3720969
#>  [3,] 0.35387594 0.7511611 0.8917204
#>  [4,] 1.06474547 1.7786540 6.4448033
#>  [5,] 2.06509209 1.4243077 5.7453426
#>  [6,] 0.07892823 1.2226107 0.2875090
#>  [7,] 1.82997763 2.3054277 0.3565721
#>  [8,] 0.23852822 0.3981631 1.9102218
#>  [9,] 0.78150857 2.9477445 0.0479901
#> [10,] 1.62660436 2.1700584 2.3323422
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.46560329 3.46560329 3.46560329
#>  [2,] 0.23651521 0.23651521 0.23651521
#>  [3,] 0.29016500 0.29016500 0.29016500
#>  [4,] 4.17398964 4.17398964 4.17398964
#>  [5,] 0.16235967 0.16235967 0.16235967
#>  [6,] 1.03818464 1.03818464 1.03818464
#>  [7,] 1.02626887 1.02626887 1.02626887
#>  [8,] 3.18682887 3.18682887 3.18682887
#>  [9,] 0.45544003 0.45544003 0.45544003
#> [10,] 0.04531748 0.04531748 0.04531748
## independence
rpextmo(10, 3, b = 1)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.2013374 0.90736751 0.7215321
#>  [2,] 0.2677459 1.11188897 1.6758983
#>  [3,] 0.2473556 0.37202142 0.8523481
#>  [4,] 1.1977503 1.38830205 0.9978927
#>  [5,] 0.3004643 2.91533127 0.1242168
#>  [6,] 0.1782324 0.01923747 3.6431074
#>  [7,] 0.1127500 1.35014486 0.6716487
#>  [8,] 3.4659888 1.32588439 0.4101442
#>  [9,] 1.0696320 0.40231643 0.3632926
#> [10,] 3.0003632 0.11374376 1.0700130

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.035421674 2.1743967 2.1743967
#>  [2,] 2.910834890 0.1153534 3.2808375
#>  [3,] 5.491822485 0.7249746 1.1715731
#>  [4,] 4.583467552 0.3102054 2.0197461
#>  [5,] 0.704875143 0.7793478 0.1775877
#>  [6,] 0.006037297 0.1461136 4.6094403
#>  [7,] 2.590829907 2.9581564 1.1550859
#>  [8,] 2.078421151 0.1387830 2.0784212
#>  [9,] 1.286822855 0.8113598 4.6166648
#> [10,] 1.926508847 1.1903645 1.5382150
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.54615050 0.54615050 0.54615050
#>  [2,] 0.32456019 0.32456019 0.32456019
#>  [3,] 0.65417299 0.65417299 0.65417299
#>  [4,] 0.84601967 0.84601967 0.84601967
#>  [5,] 1.51368117 1.51368117 1.51368117
#>  [6,] 2.28138126 2.28138126 2.28138126
#>  [7,] 0.03945684 0.03945684 0.03945684
#>  [8,] 0.12105722 0.12105722 0.12105722
#>  [9,] 0.39240031 0.39240031 0.39240031
#> [10,] 0.30353025 0.30353025 0.30353025
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.62940210 0.1994531 0.2636313
#>  [2,] 0.19820955 0.2801260 1.3319566
#>  [3,] 2.49896896 0.4048514 0.2330801
#>  [4,] 0.45812213 2.2031675 1.4174747
#>  [5,] 1.40875932 0.6382976 0.3837673
#>  [6,] 0.01516397 3.1805361 0.6714300
#>  [7,] 0.03434521 0.3507696 0.4332921
#>  [8,] 2.49472150 0.5338133 0.6744121
#>  [9,] 0.50267289 0.4167088 1.9857212
#> [10,] 1.12521143 2.3436282 1.1675091

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.82358178 0.5475557 2.6010710
#>  [2,] 0.66978569 1.6859653 1.6859653
#>  [3,] 0.66973828 0.7209740 0.7209740
#>  [4,] 0.34556166 1.0959370 0.3089304
#>  [5,] 0.16830727 0.1683073 0.1683073
#>  [6,] 3.01658821 1.4311224 0.1966993
#>  [7,] 1.75604964 4.1396045 2.1739304
#>  [8,] 0.69266790 0.6057846 2.8782797
#>  [9,] 0.09369903 0.8412174 1.8626296
#> [10,] 1.23775738 1.2377574 0.4797175
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8554062 1.8554062 1.8554062
#>  [2,] 0.0426884 0.0426884 0.0426884
#>  [3,] 0.4274064 0.4274064 0.4274064
#>  [4,] 0.5747691 0.5747691 0.5747691
#>  [5,] 0.1358138 0.1358138 0.1358138
#>  [6,] 0.3697359 0.3697359 0.3697359
#>  [7,] 1.9143141 1.9143141 1.9143141
#>  [8,] 0.6731419 0.6731419 0.6731419
#>  [9,] 0.2030975 0.2030975 0.2030975
#> [10,] 1.0451435 1.0451435 1.0451435
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.06833270 0.8093408 0.8537021
#>  [2,] 0.22135926 0.7555003 1.0065076
#>  [3,] 0.81738884 1.3021582 0.6581840
#>  [4,] 1.75139228 0.1459597 0.1345819
#>  [5,] 1.43057584 0.1138455 0.6466916
#>  [6,] 0.28789771 1.2555915 3.8089345
#>  [7,] 0.61838154 0.8172045 0.1303737
#>  [8,] 0.06468000 0.5522116 0.8164556
#>  [9,] 0.18750013 1.4502627 0.2379376
#> [10,] 0.01261314 0.2567167 0.2120806

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 3.5159964 3.5159964 0.15837109
#>  [2,] 0.3511695 0.1769257 0.44319288
#>  [3,] 0.5595504 0.6737960 0.67379597
#>  [4,] 0.1069361 0.1069361 0.10693606
#>  [5,] 0.8659352 0.8659352 0.86593524
#>  [6,] 4.3574120 1.6728966 4.35741198
#>  [7,] 2.7527293 1.6679933 0.09329555
#>  [8,] 0.2391465 0.1590190 0.14965844
#>  [9,] 3.3024308 2.9534428 0.13906871
#> [10,] 2.4893333 0.6015854 1.15953912
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.003955353 0.003955353 0.003955353
#>  [2,] 0.448660795 0.448660795 0.448660795
#>  [3,] 0.366181486 0.366181486 0.366181486
#>  [4,] 0.096965465 0.096965465 0.096965465
#>  [5,] 0.154388273 0.154388273 0.154388273
#>  [6,] 1.953766583 1.953766583 1.953766583
#>  [7,] 0.543059624 0.543059624 0.543059624
#>  [8,] 0.494944147 0.494944147 0.494944147
#>  [9,] 0.018407952 0.018407952 0.018407952
#> [10,] 2.354555195 2.354555195 2.354555195
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.28380024 1.2378625 1.27760945
#>  [2,] 1.35412230 2.4714292 0.02677620
#>  [3,] 0.95081100 0.1584311 1.27996099
#>  [4,] 0.14560397 0.4130121 1.04410043
#>  [5,] 3.63483532 0.9311667 0.11792688
#>  [6,] 1.61538569 0.1147475 0.06064094
#>  [7,] 0.03115547 1.5310856 0.41129895
#>  [8,] 0.08159151 1.1866016 0.78162176
#>  [9,] 0.62090917 0.2388383 2.37618865
#> [10,] 0.18498296 0.4478418 0.71514946

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.74011223 3.90646067 3.5494512
#>  [2,] 0.56635510 4.16626739 2.8108914
#>  [3,] 0.55521747 1.86713908 1.3402456
#>  [4,] 0.63999196 2.33003034 0.8115911
#>  [5,] 0.95609410 1.96596167 5.1023091
#>  [6,] 0.17439402 0.17439402 0.1743940
#>  [7,] 0.18750615 2.86329678 0.6637184
#>  [8,] 0.24575699 1.54041506 1.0909350
#>  [9,] 0.02527844 0.04667239 0.2274390
#> [10,] 0.07451918 0.35103555 0.3510355
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28853937 0.28853937 0.28853937
#>  [2,] 0.05630607 0.05630607 0.05630607
#>  [3,] 0.06086510 0.06086510 0.06086510
#>  [4,] 0.59597026 0.59597026 0.59597026
#>  [5,] 1.48473578 1.48473578 1.48473578
#>  [6,] 0.12482347 0.12482347 0.12482347
#>  [7,] 0.71066573 0.71066573 0.71066573
#>  [8,] 1.47603981 1.47603981 1.47603981
#>  [9,] 1.02814887 1.02814887 1.02814887
#> [10,] 1.58729166 1.58729166 1.58729166
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.13542679 1.27091153 1.1487850
#>  [2,] 0.55611617 0.11475389 0.1957381
#>  [3,] 1.29242801 1.64550936 1.6594758
#>  [4,] 0.07688704 1.12621153 2.8404270
#>  [5,] 0.23883610 2.54523431 0.9049336
#>  [6,] 1.56586453 0.68045692 1.1006982
#>  [7,] 1.06541331 0.31645218 1.3417980
#>  [8,] 1.01290906 0.83563390 0.7520568
#>  [9,] 0.38679412 1.13757823 2.3569849
#> [10,] 1.87030541 0.07027864 0.8381078

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4985967 0.4985967 0.4985967
#>  [2,] 0.1093140 0.1093140 0.1093140
#>  [3,] 0.3007675 0.9472041 0.3535235
#>  [4,] 1.1137089 1.1137089 1.1137089
#>  [5,] 4.0433850 1.0399059 2.6097371
#>  [6,] 0.2490768 0.2490768 0.2490768
#>  [7,] 3.8442236 0.4214383 0.2058646
#>  [8,] 4.1322110 5.3238310 1.3988874
#>  [9,] 3.9367615 2.9706052 0.8464447
#> [10,] 0.1085708 0.1085708 0.1085708
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40466257 0.40466257 0.40466257
#>  [2,] 1.30370837 1.30370837 1.30370837
#>  [3,] 0.44251113 0.44251113 0.44251113
#>  [4,] 0.39562265 0.39562265 0.39562265
#>  [5,] 0.22210733 0.22210733 0.22210733
#>  [6,] 0.51506723 0.51506723 0.51506723
#>  [7,] 1.08517210 1.08517210 1.08517210
#>  [8,] 0.21706257 0.21706257 0.21706257
#>  [9,] 0.87430911 0.87430911 0.87430911
#> [10,] 0.05060779 0.05060779 0.05060779
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.1014324 1.0463572 0.82410970
#>  [2,] 0.5575996 0.5584232 0.09922046
#>  [3,] 2.5819651 0.9469467 0.34795936
#>  [4,] 1.0395531 0.8196375 0.86458416
#>  [5,] 1.2209418 0.3281828 0.46180720
#>  [6,] 0.7745480 0.4864034 0.69713610
#>  [7,] 2.1817804 2.0126724 1.20451700
#>  [8,] 0.4739408 1.0869832 1.09894903
#>  [9,] 0.5182762 0.4831795 0.35389984
#> [10,] 0.4943959 0.1443353 0.05871286

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>               [,1]       [,2]       [,3]
#>  [1,] 2.7554256599 0.13349817 1.27988150
#>  [2,] 1.2361609489 0.64974918 1.93681155
#>  [3,] 0.1225448791 0.03939989 2.94350178
#>  [4,] 2.4215012993 4.33775156 1.22890768
#>  [5,] 4.0334587692 0.11850753 2.20897577
#>  [6,] 0.8473615367 0.43537810 0.45275824
#>  [7,] 0.7868645739 0.39113673 1.49156170
#>  [8,] 1.2481018342 0.48313434 3.84065501
#>  [9,] 5.6658886966 4.81557656 0.06573283
#> [10,] 0.0005825171 0.64903850 0.64903850
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26758669 0.26758669 0.26758669
#>  [2,] 0.65860812 0.65860812 0.65860812
#>  [3,] 1.26385450 1.26385450 1.26385450
#>  [4,] 0.89906109 0.89906109 0.89906109
#>  [5,] 0.16509074 0.16509074 0.16509074
#>  [6,] 0.06556968 0.06556968 0.06556968
#>  [7,] 0.29264798 0.29264798 0.29264798
#>  [8,] 0.02691698 0.02691698 0.02691698
#>  [9,] 0.14898496 0.14898496 0.14898496
#> [10,] 0.02057035 0.02057035 0.02057035
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>               [,1]       [,2]       [,3]
#>  [1,] 0.0325194624 0.29617323 0.62167629
#>  [2,] 1.1179383657 0.76075695 0.04078082
#>  [3,] 0.0004303083 3.82775678 0.57972532
#>  [4,] 0.6669760686 0.14530276 0.77099467
#>  [5,] 0.3126443031 3.54655958 0.13368418
#>  [6,] 0.2338494062 0.24662146 0.65384638
#>  [7,] 0.5539518669 0.29777237 1.28678015
#>  [8,] 0.2444695516 0.02447533 0.20620691
#>  [9,] 0.5683097354 1.84888397 0.36548724
#> [10,] 0.1581233014 2.05535363 0.64011871

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5361250 6.2145976 3.7117566
#>  [2,] 2.5689175 2.5689175 1.7236382
#>  [3,] 0.6702132 0.2909175 1.7075646
#>  [4,] 5.0971762 1.5054510 2.9701738
#>  [5,] 0.1097900 0.3566841 0.6011660
#>  [6,] 5.7325697 1.8431525 0.9277582
#>  [7,] 0.2378454 0.1875113 1.8035621
#>  [8,] 0.2046423 0.5029962 0.7509198
#>  [9,] 0.3609131 1.9422000 1.2510675
#> [10,] 0.3856110 3.3410353 0.9296146
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.664120077 0.664120077 0.664120077
#>  [2,] 0.674222047 0.674222047 0.674222047
#>  [3,] 0.536787443 0.536787443 0.536787443
#>  [4,] 0.096179419 0.096179419 0.096179419
#>  [5,] 1.476234663 1.476234663 1.476234663
#>  [6,] 3.461569115 3.461569115 3.461569115
#>  [7,] 0.306490025 0.306490025 0.306490025
#>  [8,] 0.925766997 0.925766997 0.925766997
#>  [9,] 0.081193999 0.081193999 0.081193999
#> [10,] 0.006346309 0.006346309 0.006346309
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.50838753 0.03045267 0.67145639
#>  [2,] 1.28622996 0.29633707 2.14320729
#>  [3,] 0.03559107 0.11839797 0.01030522
#>  [4,] 0.44012945 0.18413658 0.78928896
#>  [5,] 0.57742369 0.41572349 3.00439325
#>  [6,] 0.76911020 0.59008708 0.99290956
#>  [7,] 0.91582222 1.40259813 0.89564171
#>  [8,] 1.51933185 0.27494937 3.49753535
#>  [9,] 0.62534266 0.17665166 2.29414062
#> [10,] 0.37870374 0.49988102 0.13417216

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7625821 5.9025608 1.0544728
#>  [2,] 0.8775077 0.4696035 3.5875797
#>  [3,] 1.0929100 1.0929100 1.0929100
#>  [4,] 0.5888434 1.2379111 6.0577533
#>  [5,] 0.6275666 2.9710055 4.5208325
#>  [6,] 1.2140085 0.6579454 3.0962304
#>  [7,] 0.5059966 0.6843861 0.3782500
#>  [8,] 0.2107614 0.2107614 0.2107614
#>  [9,] 0.5084755 0.4419720 1.8486813
#> [10,] 0.4229653 0.1644615 0.2020169
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1532505 0.1532505 0.1532505
#>  [2,] 0.1048258 0.1048258 0.1048258
#>  [3,] 1.0327753 1.0327753 1.0327753
#>  [4,] 0.6362007 0.6362007 0.6362007
#>  [5,] 1.0668081 1.0668081 1.0668081
#>  [6,] 0.8219280 0.8219280 0.8219280
#>  [7,] 0.3259908 0.3259908 0.3259908
#>  [8,] 0.1305461 0.1305461 0.1305461
#>  [9,] 0.4348633 0.4348633 0.4348633
#> [10,] 5.9582162 5.9582162 5.9582162
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.46517597 1.9166537 0.8779960
#>  [2,] 0.67277476 1.1230486 2.2595937
#>  [3,] 1.18271700 0.8047797 2.1129500
#>  [4,] 0.03698575 1.2144003 0.3606252
#>  [5,] 0.84814904 0.1975423 1.1964477
#>  [6,] 1.02329275 2.3880762 0.1562936
#>  [7,] 0.59661633 0.1757948 0.3986040
#>  [8,] 0.54979007 0.6920786 0.3156782
#>  [9,] 2.22591250 1.4485373 0.1222401
#> [10,] 1.73110774 0.3608215 1.8390792

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4489680 8.40918411 1.0054683
#>  [2,] 1.6909612 0.26893373 0.3416920
#>  [3,] 0.4501030 0.15083056 3.7529779
#>  [4,] 3.3885812 0.06621172 1.2539411
#>  [5,] 0.3859313 0.32846901 2.9923598
#>  [6,] 0.6139919 0.94216717 3.6838540
#>  [7,] 1.0101579 0.99857157 1.0101579
#>  [8,] 0.9666303 1.20297781 0.4351703
#>  [9,] 1.4432812 2.04524594 0.1233611
#> [10,] 0.3597881 1.01449254 0.5483162
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.473156748 0.473156748 0.473156748
#>  [2,] 0.600117907 0.600117907 0.600117907
#>  [3,] 2.866999224 2.866999224 2.866999224
#>  [4,] 0.009543628 0.009543628 0.009543628
#>  [5,] 0.288185164 0.288185164 0.288185164
#>  [6,] 1.646307799 1.646307799 1.646307799
#>  [7,] 0.040009761 0.040009761 0.040009761
#>  [8,] 1.138342811 1.138342811 1.138342811
#>  [9,] 0.525698035 0.525698035 0.525698035
#> [10,] 0.690112418 0.690112418 0.690112418
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.2157826 0.45610334 0.28932573
#>  [2,] 0.6011389 0.61880909 0.01511095
#>  [3,] 1.0603369 0.15861311 0.09533958
#>  [4,] 1.3592881 0.03895054 0.56116786
#>  [5,] 2.2333965 1.29838449 0.21604537
#>  [6,] 3.8884971 0.93342139 2.80191914
#>  [7,] 0.2434813 0.88775128 0.57181424
#>  [8,] 1.8268850 0.41072183 0.73702073
#>  [9,] 0.1054028 0.81784986 1.89102790
#> [10,] 0.8975188 1.19314464 1.28886640

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.43344651 0.43344651 0.4334465
#>  [2,] 1.26215976 0.20973344 0.1301170
#>  [3,] 0.14662714 0.21401932 0.2140193
#>  [4,] 0.17540929 2.17526032 0.1754093
#>  [5,] 0.10975378 0.35969227 0.1811993
#>  [6,] 0.87628846 0.21881159 1.5656301
#>  [7,] 0.93480440 0.21134375 1.6192920
#>  [8,] 0.25938879 0.10144126 0.2593888
#>  [9,] 0.03502422 1.37758954 1.5553458
#> [10,] 0.12683921 0.01769367 0.1268392
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.31235964 0.8994616 0.08756012
#>  [2,] 0.21523959 1.0570514 0.62697308
#>  [3,] 0.15430801 0.5223251 0.16469827
#>  [4,] 1.61205565 0.3794962 0.41354363
#>  [5,] 0.77838142 1.3692197 1.36853300
#>  [6,] 0.05522538 1.4310629 1.73581056
#>  [7,] 0.09921809 0.3127431 0.31274312
#>  [8,] 1.40141462 0.8036187 1.19587956
#>  [9,] 0.84938094 0.8493809 0.83403463
#> [10,] 0.33970854 0.1908891 0.33970854
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.36875602 0.62519430 0.09457841
#>  [2,] 1.18711240 0.41623801 1.42036998
#>  [3,] 0.36628437 0.36628437 2.51668546
#>  [4,] 0.80954916 1.16697742 0.92901203
#>  [5,] 0.34923226 0.09950793 0.34923226
#>  [6,] 0.01105101 0.01105101 0.50669226
#>  [7,] 0.66323597 0.36897529 0.66323597
#>  [8,] 0.12382430 0.52100011 0.12884616
#>  [9,] 0.17741039 0.01004642 0.01004642
#> [10,] 0.82196535 2.15366407 0.85312421
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2688266 0.9594802 1.9099957
#>  [2,] 0.4032997 0.3940730 0.5571534
#>  [3,] 1.2397923 0.2446990 0.1175876
#>  [4,] 1.0516345 0.5810021 0.8384738
#>  [5,] 0.1817349 0.4307516 1.3184454
#>  [6,] 0.5743677 0.6812961 0.3993433
#>  [7,] 0.1399801 0.1399801 0.1399801
#>  [8,] 0.5756548 0.4223074 1.5977840
#>  [9,] 0.7644051 2.5281738 1.4073157
#> [10,] 0.1175664 0.8397040 0.1175664
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35224521 0.35224521 0.02965799
#>  [2,] 0.60717613 3.39066368 1.28682021
#>  [3,] 0.50901973 0.50901973 2.13872665
#>  [4,] 1.10120040 3.26057043 1.58443986
#>  [5,] 0.55789622 0.19289587 0.19289587
#>  [6,] 0.03813749 0.03813749 1.19115182
#>  [7,] 0.03584456 1.26379416 0.03584456
#>  [8,] 0.45859827 0.97511437 0.35263728
#>  [9,] 0.27976775 1.55924869 1.55924869
#> [10,] 0.25206443 0.46776705 0.46776705

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.30906041 0.3090604 0.30906041
#>  [2,] 0.50548515 4.3005901 2.71981230
#>  [3,] 1.10925652 0.5181449 0.40241811
#>  [4,] 0.57987387 0.7342224 1.30189855
#>  [5,] 0.49894092 2.7352221 0.80227997
#>  [6,] 0.44866384 0.4486638 0.44866384
#>  [7,] 0.18348873 0.6328819 0.63288195
#>  [8,] 0.53275023 0.5381972 0.09233334
#>  [9,] 1.83531343 0.3074954 3.60756519
#> [10,] 0.02859069 0.1150357 0.68224930
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6525999 1.4059180 0.7620753
#>  [2,] 1.7232350 1.6270613 1.7232350
#>  [3,] 1.1570671 0.8516672 0.3046248
#>  [4,] 1.1178887 4.0133679 1.6325652
#>  [5,] 0.8018310 1.6834416 2.1077709
#>  [6,] 0.5379535 0.5379535 0.5379535
#>  [7,] 1.7267064 0.2047581 2.8357981
#>  [8,] 1.0350719 0.5204165 0.9071268
#>  [9,] 0.8970036 0.2636169 1.5581745
#> [10,] 0.7607842 0.7607842 0.7607842
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.49979100 1.93643264 0.2395464
#>  [2,] 0.09210135 0.02672511 1.2827071
#>  [3,] 0.06121278 0.06121278 1.7229999
#>  [4,] 2.02937634 0.33581232 2.0293763
#>  [5,] 1.89600348 1.70018338 1.6815628
#>  [6,] 1.58374635 2.06492221 0.6014179
#>  [7,] 3.09501085 1.37901387 0.7487753
#>  [8,] 0.15093191 0.15093191 0.1509319
#>  [9,] 3.70313350 2.97299803 2.7791299
#> [10,] 3.75405826 3.75405826 2.4550096
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.7617830 1.66595715 1.76178297
#>  [2,] 1.5009648 2.44576002 0.88356144
#>  [3,] 0.9784880 0.97848799 0.97848799
#>  [4,] 0.4118808 0.87303262 0.61732326
#>  [5,] 0.2279721 0.22797212 0.22797212
#>  [6,] 0.8179487 0.81794872 0.81794872
#>  [7,] 0.4268349 0.07160987 0.58531171
#>  [8,] 0.7960021 3.76654870 0.03479496
#>  [9,] 3.6224358 1.07060380 0.79735813
#> [10,] 0.8133352 0.81333517 0.81333517
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>            [,1]      [,2]        [,3]
#>  [1,] 3.6534560 3.0403129 3.063177572
#>  [2,] 0.6923366 0.2406506 3.201325853
#>  [3,] 0.8778008 0.8778008 0.877800794
#>  [4,] 1.8063044 0.5170527 0.128674275
#>  [5,] 3.3521632 0.6957270 2.125327317
#>  [6,] 0.8945111 0.8945111 0.849829743
#>  [7,] 0.9566483 0.9566483 0.956648259
#>  [8,] 1.4128044 1.4128044 0.008107269
#>  [9,] 2.0754083 3.3602187 2.918430850
#> [10,] 1.4801209 1.5242818 1.524281801

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.06903730 0.007216985 0.06903730
#>  [2,] 0.34345487 0.560039028 0.24223999
#>  [3,] 0.12180727 0.121807273 0.12180727
#>  [4,] 1.64163998 1.720842006 0.28534858
#>  [5,] 0.62788306 0.984809185 1.88771315
#>  [6,] 1.04624073 1.722018012 0.49285530
#>  [7,] 0.84201478 0.842014782 0.54714423
#>  [8,] 1.26952291 1.269522907 0.53213701
#>  [9,] 0.04017785 0.860330126 0.04017785
#> [10,] 0.21405866 0.214058658 0.21405866
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36209078 0.52987819 0.52987819
#>  [2,] 0.07880149 0.07880149 0.07880149
#>  [3,] 3.42645748 0.13869266 0.57398870
#>  [4,] 0.04575070 0.15674329 0.26603583
#>  [5,] 0.46858603 0.46858603 0.46858603
#>  [6,] 0.37409185 0.37409185 0.17098775
#>  [7,] 0.46507008 0.46507008 0.46507008
#>  [8,] 0.28664395 0.28664395 0.28664395
#>  [9,] 0.38490567 0.38490567 0.38490567
#> [10,] 0.62660157 1.31699440 1.31699440
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.94985196 1.49425978 0.94985196
#>  [2,] 0.06588256 0.27430492 0.27430492
#>  [3,] 0.01181365 0.08316825 0.08316825
#>  [4,] 1.27942299 0.44498418 1.11727938
#>  [5,] 0.26718074 0.24591213 0.26718074
#>  [6,] 0.01748197 0.01748197 1.54610223
#>  [7,] 2.35003339 2.73710932 1.03888779
#>  [8,] 0.83310360 0.83310360 0.83310360
#>  [9,] 0.03729808 0.03729808 0.22871967
#> [10,] 0.11765777 0.88040682 1.32422453
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04316921 0.04316921 0.04316921
#>  [2,] 0.36992838 0.36992838 0.36992838
#>  [3,] 0.68949926 0.19404253 0.19404253
#>  [4,] 0.53267972 0.43739443 0.53267972
#>  [5,] 0.10355845 0.10355845 0.10355845
#>  [6,] 0.92977887 0.92977887 0.39025991
#>  [7,] 1.75057427 1.27774636 1.20685815
#>  [8,] 0.01757906 0.68371618 0.47181535
#>  [9,] 0.17466661 0.17466661 0.08452283
#> [10,] 0.25884555 0.25884555 0.25884555
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.01011061 0.1473390 0.1473390
#>  [2,] 0.58587749 0.8470564 1.0111326
#>  [3,] 1.20697563 0.9730524 0.9038394
#>  [4,] 0.25751866 0.2575187 0.2575187
#>  [5,] 0.19310999 0.1931100 0.1931100
#>  [6,] 1.79258753 0.1936062 1.7925875
#>  [7,] 1.38649536 1.3864954 1.3864954
#>  [8,] 0.10406324 0.1040632 0.0600808
#>  [9,] 0.61307836 0.6130784 0.6130784
#> [10,] 0.08982651 1.1995923 2.3445008

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.05337902 0.05337902 0.053379023
#>  [2,] 0.11646652 0.26036119 0.641106576
#>  [3,] 0.20774551 0.20774551 0.207745512
#>  [4,] 0.24904865 0.32531027 0.087760906
#>  [5,] 0.16864510 0.81701546 0.817015462
#>  [6,] 0.09876059 0.09876059 0.098760588
#>  [7,] 0.27043007 0.19225655 0.270430072
#>  [8,] 0.59185677 0.11892700 0.591856767
#>  [9,] 0.21789554 0.21789554 0.107489850
#> [10,] 0.86814811 1.13646004 0.003154008
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.29177524 0.3559152 0.64573293
#>  [2,] 0.78900142 0.7890014 0.29241174
#>  [3,] 0.06568721 0.2294477 0.33576446
#>  [4,] 1.14104605 1.0795891 0.07982999
#>  [5,] 0.44925179 0.4871189 0.48711889
#>  [6,] 0.55830411 0.5583041 0.17560119
#>  [7,] 0.62688299 0.8362281 1.47358442
#>  [8,] 0.21594684 0.2159468 0.21594684
#>  [9,] 0.03202388 0.1931312 0.13154337
#> [10,] 0.25550657 0.4302133 0.29332542
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.54492895 0.54492895 0.54492895
#>  [2,] 0.69837333 0.17782349 0.40627894
#>  [3,] 0.01174214 0.01174214 0.01174214
#>  [4,] 0.19549858 0.21225916 0.37286109
#>  [5,] 0.10861857 0.10861857 0.10861857
#>  [6,] 0.18681502 0.18681502 0.18681502
#>  [7,] 0.78057806 0.29264465 0.76761254
#>  [8,] 0.86660162 0.86660162 0.54704027
#>  [9,] 0.13566568 0.03952707 0.67931567
#> [10,] 1.43061170 0.48072065 0.52999992
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 0.1758089 0.872777744 0.56230498
#>  [2,] 0.1586458 0.350718322 0.21382619
#>  [3,] 0.2056796 0.392467209 0.08429771
#>  [4,] 0.6246188 0.544505051 0.60285490
#>  [5,] 0.2220677 0.222067696 0.22176544
#>  [6,] 0.3712765 0.371276479 0.37127648
#>  [7,] 0.4487334 0.005140756 0.06901422
#>  [8,] 0.4906617 0.914818493 0.46467061
#>  [9,] 0.3916599 0.289607085 0.88745852
#> [10,] 0.1385418 0.138541780 0.13854178

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40135853 1.11773434 0.40607601
#>  [2,] 0.78484481 0.97005285 0.70563864
#>  [3,] 0.07794055 0.07794055 0.53874925
#>  [4,] 0.50754400 0.50754400 0.16237621
#>  [5,] 0.43087157 0.14725802 1.30602100
#>  [6,] 0.55028622 0.04256847 0.55028622
#>  [7,] 0.47770865 0.47770865 0.48450218
#>  [8,] 0.26484911 0.12411338 0.26494738
#>  [9,] 0.05809661 0.13715634 0.24479789
#> [10,] 0.57445107 0.37410269 0.09488846
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 0.009496422 0.06745460 0.067454604
#>  [2,] 0.112688328 0.19859216 0.198592162
#>  [3,] 0.223339157 0.22333916 0.223339157
#>  [4,] 0.428285249 0.23527312 0.235273116
#>  [5,] 0.252657652 0.19929447 0.028124047
#>  [6,] 0.005213461 0.95495559 0.005213461
#>  [7,] 0.486385730 0.03989582 0.039895824
#>  [8,] 0.179944124 0.40019590 0.116600683
#>  [9,] 0.011692505 0.01169250 0.011692505
#> [10,] 0.587070887 0.16247474 0.482739716
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.30667621 0.24653627 0.75128252
#>  [2,] 0.53990118 0.25101905 0.04725444
#>  [3,] 0.06159947 0.10282099 0.06159947
#>  [4,] 0.54389980 0.54389980 0.54389980
#>  [5,] 0.42368036 0.44620209 0.33209900
#>  [6,] 0.16221222 0.32083802 0.16221222
#>  [7,] 1.17150433 0.03274080 0.66208655
#>  [8,] 0.69387488 0.26493928 0.29762238
#>  [9,] 0.94358815 0.05408558 0.99852104
#> [10,] 0.31417241 0.31417241 0.31417241
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.78557634 0.23269608 0.23269608
#>  [2,] 0.53669159 0.04529631 0.01327192
#>  [3,] 0.12034028 0.12034028 0.38149845
#>  [4,] 1.08043001 1.19807916 1.20117502
#>  [5,] 0.80502981 1.12751514 0.17304689
#>  [6,] 0.53462132 0.41938719 0.65414547
#>  [7,] 0.01150003 0.01150003 0.01150003
#>  [8,] 0.34689428 0.13030256 0.34689428
#>  [9,] 0.54393735 0.39827981 0.54542854
#> [10,] 0.01020722 0.03304975 0.18341268

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>              [,1]        [,2]       [,3]
#>  [1,] 0.003514077 0.003514077 0.37858712
#>  [2,] 0.089895425 0.217536101 0.09660158
#>  [3,] 0.128588021 0.128588021 0.12858802
#>  [4,] 0.465672110 0.465672110 0.46567211
#>  [5,] 0.111897735 0.033256905 0.11189774
#>  [6,] 0.421851656 0.544902991 1.02501647
#>  [7,] 0.121407034 0.121407034 0.12140703
#>  [8,] 0.492661196 0.492661196 0.49266120
#>  [9,] 0.228403330 0.069979361 0.22840333
#> [10,] 0.282868659 0.649964792 0.64996479
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.16141184 0.16141184 0.16141184
#>  [2,] 0.96326427 0.96326427 0.96326427
#>  [3,] 0.06008818 0.06008818 0.06008818
#>  [4,] 0.20639325 0.34676066 0.54890052
#>  [5,] 0.19356253 0.47011962 0.36854839
#>  [6,] 0.23881659 0.56954415 0.50771714
#>  [7,] 0.30764939 0.40955347 0.30764939
#>  [8,] 0.43950624 0.00646038 0.43950624
#>  [9,] 0.33988782 0.07230026 0.33988782
#> [10,] 0.21985952 0.03452843 0.15466310
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.386207462 0.386207462 0.386207462
#>  [2,] 0.006695344 0.001053115 0.006695344
#>  [3,] 1.302809643 1.302809643 0.816972523
#>  [4,] 1.193442274 0.537296839 1.193442274
#>  [5,] 0.014754570 0.014754570 0.014754570
#>  [6,] 0.381726549 0.381726549 0.381726549
#>  [7,] 0.419573970 0.426515287 0.419573970
#>  [8,] 0.538132314 0.079544846 0.828832943
#>  [9,] 0.197453981 0.479268549 0.197453981
#> [10,] 0.035848728 0.035848728 0.035848728
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.75171180 0.75171180 0.55565489
#>  [2,] 0.12298233 0.12298233 0.21102365
#>  [3,] 0.42996862 0.16419774 0.42996862
#>  [4,] 0.39705806 0.39705806 1.12371941
#>  [5,] 0.05574795 0.05569687 0.05569687
#>  [6,] 0.04237964 0.89420253 0.01514909
#>  [7,] 0.01277930 0.01277930 0.01277930
#>  [8,] 0.59922661 0.67146803 0.09621333
#>  [9,] 0.12693545 0.49460209 0.39569402
#> [10,] 0.72085994 0.74841052 0.74841052