Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.5105572 0.25054623 0.90112323
#>  [2,] 0.8864402 0.88644021 0.88644021
#>  [3,] 0.3328203 0.89086952 0.01181449
#>  [4,] 1.1906121 0.43156777 0.40770803
#>  [5,] 4.7454255 4.70169376 3.08019930
#>  [6,] 1.1693087 2.23892690 2.23892690
#>  [7,] 1.1360672 2.58676282 1.26717748
#>  [8,] 0.3898011 0.38980108 0.38980108
#>  [9,] 3.9256287 0.08513026 0.68043132
#> [10,] 2.9062292 0.67335266 2.28309582
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18686143 0.18686143 0.18686143
#>  [2,] 0.22221006 0.22221006 0.22221006
#>  [3,] 1.04838936 1.04838936 1.04838936
#>  [4,] 0.50259987 0.50259987 0.50259987
#>  [5,] 0.09531318 0.09531318 0.09531318
#>  [6,] 0.36173716 0.36173716 0.36173716
#>  [7,] 1.12900821 1.12900821 1.12900821
#>  [8,] 0.09876781 0.09876781 0.09876781
#>  [9,] 1.81114755 1.81114755 1.81114755
#> [10,] 0.49282021 0.49282021 0.49282021
## independence
rpextmo(10, 3, b = 1)
#>            [,1]         [,2]       [,3]
#>  [1,] 0.4994646 0.2589009246 0.12085336
#>  [2,] 0.1111108 0.5549015091 0.69940965
#>  [3,] 2.0708434 0.6775319888 0.07154416
#>  [4,] 0.8752295 0.0347608263 0.46238638
#>  [5,] 0.1852044 0.4185318681 1.05450934
#>  [6,] 0.8135621 0.0008391423 0.17605567
#>  [7,] 0.5746443 1.3907080813 0.08654230
#>  [8,] 1.9873802 0.8204131564 1.96281016
#>  [9,] 0.4135371 0.4729689779 0.35725243
#> [10,] 0.1012517 2.2353021164 0.06031046

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.28807284 0.3589017 1.0122348
#>  [2,] 0.98473887 0.6090767 1.6519721
#>  [3,] 0.51237177 1.6948974 0.3766718
#>  [4,] 2.25121280 0.5011450 2.2512128
#>  [5,] 1.96293628 1.9629363 0.1720330
#>  [6,] 0.78236180 1.8502704 0.2328692
#>  [7,] 2.65099120 0.5777366 6.1844086
#>  [8,] 0.04873126 2.1898016 0.3741137
#>  [9,] 0.23541060 0.5412400 0.5412400
#> [10,] 2.50997099 2.5099710 1.5187561
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7826626 0.7826626 0.7826626
#>  [2,] 0.4313829 0.4313829 0.4313829
#>  [3,] 1.0160116 1.0160116 1.0160116
#>  [4,] 1.2263517 1.2263517 1.2263517
#>  [5,] 0.2330040 0.2330040 0.2330040
#>  [6,] 0.8683546 0.8683546 0.8683546
#>  [7,] 0.1354794 0.1354794 0.1354794
#>  [8,] 0.1746091 0.1746091 0.1746091
#>  [9,] 0.4120919 0.4120919 0.4120919
#> [10,] 0.6563704 0.6563704 0.6563704
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1409397 0.07862869 1.0757417
#>  [2,] 0.3595047 0.15243139 1.4326522
#>  [3,] 0.4895675 0.01692646 0.8830197
#>  [4,] 0.4381061 0.42643890 0.7402994
#>  [5,] 1.8344993 0.43798449 0.1409267
#>  [6,] 1.2445759 0.67007716 0.4626504
#>  [7,] 0.3925604 0.74196756 1.6440246
#>  [8,] 1.7500711 0.16618259 0.8574319
#>  [9,] 1.9519170 0.16018574 1.8302505
#> [10,] 0.1920190 2.68493767 0.6283209

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.47423235 0.2637517 0.97059656
#>  [2,] 2.96286007 3.8053847 4.11208372
#>  [3,] 1.60338376 0.1446251 0.08771165
#>  [4,] 0.08792937 0.4612575 1.18398195
#>  [5,] 3.65821808 3.8924917 2.76054879
#>  [6,] 1.27881817 1.5917956 0.48067984
#>  [7,] 3.68144927 1.8555117 0.46720786
#>  [8,] 0.95187269 1.7752136 1.77521359
#>  [9,] 1.01714694 2.0028088 1.85796715
#> [10,] 1.70891508 2.2811353 2.73723498
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2549560 0.2549560 0.2549560
#>  [2,] 0.2516215 0.2516215 0.2516215
#>  [3,] 1.0063459 1.0063459 1.0063459
#>  [4,] 1.1732428 1.1732428 1.1732428
#>  [5,] 0.5900847 0.5900847 0.5900847
#>  [6,] 1.3852717 1.3852717 1.3852717
#>  [7,] 0.3818325 0.3818325 0.3818325
#>  [8,] 0.6497955 0.6497955 0.6497955
#>  [9,] 1.3321338 1.3321338 1.3321338
#> [10,] 1.0358970 1.0358970 1.0358970
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.18075483 0.53128425 1.4138308
#>  [2,] 2.93818695 0.30048277 0.8445162
#>  [3,] 0.53921874 2.95277747 1.4274605
#>  [4,] 0.01474536 0.02570415 1.3912712
#>  [5,] 0.64309114 2.86148741 1.8633284
#>  [6,] 1.51923575 0.47675493 0.4043051
#>  [7,] 3.03150089 0.01268266 0.2979085
#>  [8,] 0.49610828 0.32846806 2.3959994
#>  [9,] 0.48478537 0.79505811 0.6572990
#> [10,] 2.55423243 0.07709955 0.4652713

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 4.4269077 0.3846390 0.87801788
#>  [2,] 0.1798212 5.0461857 1.04597705
#>  [3,] 0.6029984 2.4868593 2.48685927
#>  [4,] 0.3928198 0.3928198 0.01630195
#>  [5,] 1.2861472 2.6868847 0.40515591
#>  [6,] 1.9541348 4.4878796 0.76709602
#>  [7,] 0.2374638 0.6632064 0.22648744
#>  [8,] 0.5316576 2.1217608 2.04287374
#>  [9,] 1.5781016 0.8938686 2.63227558
#> [10,] 2.7512878 0.4055228 0.86875995
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1042745 1.1042745 1.1042745
#>  [2,] 0.3293326 0.3293326 0.3293326
#>  [3,] 0.3562873 0.3562873 0.3562873
#>  [4,] 1.2819419 1.2819419 1.2819419
#>  [5,] 0.9094795 0.9094795 0.9094795
#>  [6,] 0.5691433 0.5691433 0.5691433
#>  [7,] 0.7716583 0.7716583 0.7716583
#>  [8,] 0.1877147 0.1877147 0.1877147
#>  [9,] 0.2285006 0.2285006 0.2285006
#> [10,] 0.8900008 0.8900008 0.8900008
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.70080687 0.08922936 1.3213873
#>  [2,] 0.98458058 0.51510409 1.2672406
#>  [3,] 0.09848434 0.46281647 1.5667652
#>  [4,] 0.84537111 1.43337181 1.3482338
#>  [5,] 1.42688907 0.61817596 0.1424997
#>  [6,] 3.92688356 0.29201585 2.5828394
#>  [7,] 0.92668290 2.12630695 0.9144841
#>  [8,] 2.02424115 0.81769748 0.7476296
#>  [9,] 0.57234201 0.32244087 0.8193467
#> [10,] 0.04169248 0.33000194 1.2454849

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.96230071 2.3125648 3.63610216
#>  [2,] 2.11150964 1.0766236 0.01070794
#>  [3,] 1.84513197 2.9110093 0.56286012
#>  [4,] 0.04316037 0.1355325 4.03810601
#>  [5,] 1.05284895 0.1001034 2.82956595
#>  [6,] 1.29604614 1.0015347 6.02942786
#>  [7,] 0.02464461 2.5014654 3.63469503
#>  [8,] 0.98575190 1.0496002 1.19867128
#>  [9,] 0.47993589 0.4022491 2.58444676
#> [10,] 0.67685680 5.2601212 0.15613644
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0480089 1.0480089 1.0480089
#>  [2,] 2.3018521 2.3018521 2.3018521
#>  [3,] 1.2184531 1.2184531 1.2184531
#>  [4,] 0.1242884 0.1242884 0.1242884
#>  [5,] 0.2025275 0.2025275 0.2025275
#>  [6,] 0.6790574 0.6790574 0.6790574
#>  [7,] 0.5439820 0.5439820 0.5439820
#>  [8,] 0.4243739 0.4243739 0.4243739
#>  [9,] 0.7113938 0.7113938 0.7113938
#> [10,] 2.8107950 2.8107950 2.8107950
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1262349 0.3982683 1.2582796
#>  [2,] 4.8154105 0.9424230 0.4089376
#>  [3,] 2.7409481 3.9538278 0.8402452
#>  [4,] 0.0317372 0.2514146 0.9409279
#>  [5,] 0.1157481 1.2765222 0.6528058
#>  [6,] 0.5280489 0.8034038 0.3960849
#>  [7,] 0.6438369 0.1327380 0.7431283
#>  [8,] 1.8802544 1.9766507 1.3296360
#>  [9,] 0.1320565 0.6060908 0.2987621
#> [10,] 2.4801481 0.7845605 0.2058714

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.12157483 0.9164809 0.5257023
#>  [2,] 2.82669594 0.9328008 2.7468678
#>  [3,] 0.71268001 2.8814870 0.5456505
#>  [4,] 3.03856796 1.1370349 4.6822342
#>  [5,] 0.39526209 0.9414484 0.1527604
#>  [6,] 0.27645639 0.2764564 0.2764564
#>  [7,] 0.02725177 1.2767473 2.0865534
#>  [8,] 2.09035762 2.0091284 2.6746305
#>  [9,] 0.03317224 1.4285649 0.2523741
#> [10,] 0.99504686 0.2599692 0.2446995
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.68240289 2.68240289 2.68240289
#>  [2,] 2.29693077 2.29693077 2.29693077
#>  [3,] 0.77664144 0.77664144 0.77664144
#>  [4,] 1.26133329 1.26133329 1.26133329
#>  [5,] 4.53269963 4.53269963 4.53269963
#>  [6,] 2.10262536 2.10262536 2.10262536
#>  [7,] 0.26587298 0.26587298 0.26587298
#>  [8,] 0.10029844 0.10029844 0.10029844
#>  [9,] 0.08630974 0.08630974 0.08630974
#> [10,] 0.30968022 0.30968022 0.30968022
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.25625728 1.21610474 0.52321248
#>  [2,] 0.47030289 1.60233959 0.05463272
#>  [3,] 0.81551905 0.30316813 1.30269742
#>  [4,] 1.25248596 4.43618809 0.33111952
#>  [5,] 1.11572451 1.37395942 1.65650088
#>  [6,] 0.67342724 0.03347572 0.53343972
#>  [7,] 1.02618575 2.06054089 0.52367630
#>  [8,] 2.16189347 0.14259077 0.28365745
#>  [9,] 0.88358183 0.49218399 1.69171986
#> [10,] 0.09201863 0.79085765 2.06354690

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 3.06568951 3.0260658 1.8943893
#>  [2,] 1.91575175 0.1057741 3.2039968
#>  [3,] 0.61249988 0.6124999 0.2215732
#>  [4,] 0.67476102 0.4728821 3.3074283
#>  [5,] 0.69120257 0.5158391 0.1066075
#>  [6,] 1.00815917 0.1525639 7.7434539
#>  [7,] 0.59556048 0.2092601 0.5955605
#>  [8,] 0.04299858 2.3939894 0.5936410
#>  [9,] 1.85313463 2.0772971 0.2434977
#> [10,] 4.13159626 2.3212357 2.5888475
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.2818540 2.2818540 2.2818540
#>  [2,] 0.2577470 0.2577470 0.2577470
#>  [3,] 1.1022822 1.1022822 1.1022822
#>  [4,] 0.7780158 0.7780158 0.7780158
#>  [5,] 2.0108200 2.0108200 2.0108200
#>  [6,] 3.2378281 3.2378281 3.2378281
#>  [7,] 1.0615327 1.0615327 1.0615327
#>  [8,] 0.4489321 0.4489321 0.4489321
#>  [9,] 0.1082202 0.1082202 0.1082202
#> [10,] 0.1999944 0.1999944 0.1999944
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4209345 2.00023318 0.11442622
#>  [2,] 0.6076814 0.52285784 0.47542684
#>  [3,] 2.7103257 1.41767257 1.96174147
#>  [4,] 0.3380842 0.50653081 0.53080158
#>  [5,] 0.9348789 1.23247834 0.08683222
#>  [6,] 1.2137855 0.06557269 0.92702188
#>  [7,] 0.6228151 0.51472914 0.29929788
#>  [8,] 2.0503365 3.03521172 0.87225776
#>  [9,] 1.1734251 1.32822941 0.47138553
#> [10,] 0.3493396 0.24988336 3.26154634

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.8370736 0.3590809 3.8580637
#>  [2,] 0.9106737 0.9106737 0.3924151
#>  [3,] 3.0555505 0.7014344 3.6793486
#>  [4,] 0.6537544 0.6537544 0.6537544
#>  [5,] 0.2489657 0.2333723 1.5397422
#>  [6,] 0.4799498 3.4730882 5.6417769
#>  [7,] 2.7899240 3.9286700 0.6718798
#>  [8,] 0.3090980 3.1296726 1.3908292
#>  [9,] 1.0783850 6.3248789 1.9264168
#> [10,] 0.1726712 1.3529075 1.3529075
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8672486 0.8672486 0.8672486
#>  [2,] 0.3481455 0.3481455 0.3481455
#>  [3,] 1.8857604 1.8857604 1.8857604
#>  [4,] 1.3223756 1.3223756 1.3223756
#>  [5,] 0.7770837 0.7770837 0.7770837
#>  [6,] 0.3612805 0.3612805 0.3612805
#>  [7,] 0.2703690 0.2703690 0.2703690
#>  [8,] 2.1119786 2.1119786 2.1119786
#>  [9,] 0.4833439 0.4833439 0.4833439
#> [10,] 0.4112948 0.4112948 0.4112948
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]        [,3]
#>  [1,] 1.3993795 2.2585549 0.908418005
#>  [2,] 1.5647662 1.9173484 0.686006236
#>  [3,] 0.1207894 0.3499520 0.053519480
#>  [4,] 1.6572174 0.1395942 1.349975693
#>  [5,] 0.6394535 3.1244099 0.001813297
#>  [6,] 0.3148191 1.3572893 0.062575516
#>  [7,] 1.8738131 2.0969567 0.193207601
#>  [8,] 1.6286787 0.3606483 0.574830594
#>  [9,] 1.7597805 1.2319237 0.019713207
#> [10,] 1.7479852 0.8153479 0.511177592

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0778160 3.2085485 2.4978750
#>  [2,] 0.2389050 0.4541405 0.9525051
#>  [3,] 0.3823464 0.3823464 0.3823464
#>  [4,] 0.5689343 0.5689343 0.5689343
#>  [5,] 0.4603709 0.5946163 0.7357387
#>  [6,] 0.2871179 0.5038284 3.5218275
#>  [7,] 2.3110806 0.4504383 4.2318121
#>  [8,] 0.6986897 0.6986897 0.6986897
#>  [9,] 2.2694270 0.2804226 3.1448442
#> [10,] 0.6667094 1.3756056 0.4091591
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.48255577 0.48255577 0.48255577
#>  [2,] 0.11792967 0.11792967 0.11792967
#>  [3,] 0.64625152 0.64625152 0.64625152
#>  [4,] 0.86935412 0.86935412 0.86935412
#>  [5,] 0.07192411 0.07192411 0.07192411
#>  [6,] 0.14502713 0.14502713 0.14502713
#>  [7,] 1.23793790 1.23793790 1.23793790
#>  [8,] 2.38690222 2.38690222 2.38690222
#>  [9,] 0.10659089 0.10659089 0.10659089
#> [10,] 0.94120857 0.94120857 0.94120857
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 4.44568135 1.1790284 6.4559218
#>  [2,] 1.93064296 0.8164702 0.4074837
#>  [3,] 4.18000089 0.4725159 0.1918128
#>  [4,] 0.39295149 0.1203332 3.2944715
#>  [5,] 1.06919893 0.4428527 0.7306253
#>  [6,] 0.53994161 1.2208772 2.4479872
#>  [7,] 1.50824460 2.6505749 0.7286123
#>  [8,] 0.79984433 0.5857773 0.1362660
#>  [9,] 0.01271572 2.4696533 4.7603838
#> [10,] 1.48485804 1.4544256 1.0085449

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.5510860 1.5466984 2.3353831
#>  [2,] 6.8752855 1.4641312 1.8675973
#>  [3,] 0.2282630 0.1414797 2.4219823
#>  [4,] 0.2491009 0.2491009 0.2491009
#>  [5,] 0.3359499 0.4664418 0.4664418
#>  [6,] 3.1490968 2.2743401 3.7446517
#>  [7,] 1.0679431 1.0679431 1.0679431
#>  [8,] 0.3397652 0.3397652 0.3397652
#>  [9,] 1.1261996 0.2293466 1.3125458
#> [10,] 2.4626283 1.2149925 0.2681656
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23223780 0.23223780 0.23223780
#>  [2,] 1.95023798 1.95023798 1.95023798
#>  [3,] 2.09404738 2.09404738 2.09404738
#>  [4,] 0.59374728 0.59374728 0.59374728
#>  [5,] 0.13823249 0.13823249 0.13823249
#>  [6,] 0.52594382 0.52594382 0.52594382
#>  [7,] 0.45245415 0.45245415 0.45245415
#>  [8,] 0.06655585 0.06655585 0.06655585
#>  [9,] 1.84130606 1.84130606 1.84130606
#> [10,] 0.67836411 0.67836411 0.67836411
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.48676507 1.85320732 4.2911326
#>  [2,] 0.56881389 1.30873697 0.7844867
#>  [3,] 0.03606975 2.04987899 1.2881732
#>  [4,] 1.40553021 0.91509817 0.3596181
#>  [5,] 0.61533519 1.20641501 0.6483876
#>  [6,] 4.45677702 0.89224359 4.0817586
#>  [7,] 2.41988112 0.12945438 0.5214997
#>  [8,] 0.41540907 0.14501707 2.3955727
#>  [9,] 1.46735055 4.74550641 3.4217828
#> [10,] 0.05759898 0.06945602 1.4327387

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.27175184 0.1547037 0.27175184
#>  [2,] 3.09054210 1.3223588 0.01784441
#>  [3,] 2.75735315 2.1323998 0.01941866
#>  [4,] 0.99549874 0.3434270 0.47956588
#>  [5,] 0.15215772 0.6384575 2.24131442
#>  [6,] 0.67944275 1.7239730 0.67944275
#>  [7,] 0.03587504 0.2104920 0.08405202
#>  [8,] 2.47517463 2.3565567 0.88889393
#>  [9,] 0.62962888 0.9255207 0.92552070
#> [10,] 0.83383270 0.6495795 0.04353398
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>             [,1]      [,2]         [,3]
#>  [1,] 0.04939245 1.2352311 0.0350851971
#>  [2,] 0.38743415 0.3874341 0.3874341490
#>  [3,] 0.02898456 0.3497088 0.0006346966
#>  [4,] 4.99589662 0.3545414 0.3534877666
#>  [5,] 0.20849914 0.2084991 0.2084991389
#>  [6,] 0.79474367 0.2737295 0.7358021054
#>  [7,] 0.95450065 0.4375042 0.0250896946
#>  [8,] 1.77469054 0.3170249 0.7776366639
#>  [9,] 0.96509763 0.9650976 0.0972281508
#> [10,] 0.82656292 2.1491145 0.4288458696
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.6423207 0.9605561 0.21963567
#>  [2,] 0.1973036 1.0743510 0.06551236
#>  [3,] 0.9436428 0.1128041 0.41167686
#>  [4,] 2.1371183 0.4719508 1.27078806
#>  [5,] 0.6420728 4.4076798 3.87203338
#>  [6,] 1.3594145 0.1916224 0.19162241
#>  [7,] 0.7326568 1.0074778 0.25900681
#>  [8,] 0.5824278 0.4852901 0.95304642
#>  [9,] 1.9711267 0.8326914 1.44167735
#> [10,] 0.6506095 0.7270563 0.90759889
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.58157910 0.5198648 0.4444454
#>  [2,] 0.20712550 0.2071255 0.2071255
#>  [3,] 0.84604244 0.6637855 0.8460424
#>  [4,] 1.34143844 1.1549083 0.1664876
#>  [5,] 0.28171710 0.4734616 0.4734616
#>  [6,] 0.31686700 0.2370203 0.9042128
#>  [7,] 0.02961196 0.2661954 0.9535778
#>  [8,] 0.48860338 1.2995406 1.0883716
#>  [9,] 0.33554490 0.7202584 1.0349747
#> [10,] 0.16958129 0.1695813 0.4018378
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.394986873 1.343272340 0.34968714
#>  [2,] 0.108294490 0.741224338 0.04191658
#>  [3,] 0.320803608 0.181538729 0.23362932
#>  [4,] 0.064960237 0.064960237 0.65165972
#>  [5,] 0.312978638 0.312978638 0.31297864
#>  [6,] 0.832786969 0.005463576 0.77662324
#>  [7,] 0.079729671 0.079729671 0.07972967
#>  [8,] 0.622889476 1.832137909 0.93246420
#>  [9,] 0.009797443 0.869255875 0.08617379
#> [10,] 0.054599058 1.590288634 1.11144610

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.7487944 0.74879439 0.04282567
#>  [2,] 0.5898678 0.58986778 0.58986778
#>  [3,] 1.6169309 0.51105078 4.08412837
#>  [4,] 0.9269367 0.92693672 0.72721334
#>  [5,] 2.5466746 0.06127845 2.54667463
#>  [6,] 0.1332195 1.64455676 3.82756164
#>  [7,] 3.3425970 3.43322912 1.54063928
#>  [8,] 1.4719257 1.15034672 1.63695996
#>  [9,] 1.1236187 0.86373424 1.12361872
#> [10,] 1.5876961 1.74367731 1.74367731
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6377271 0.8018587 3.1000282
#>  [2,] 1.1661473 1.6217604 1.6217604
#>  [3,] 0.7223802 4.8755221 1.5345038
#>  [4,] 2.2289315 0.5072204 0.6031696
#>  [5,] 3.0748458 3.4162741 0.8675170
#>  [6,] 5.4442993 5.0664297 4.8835030
#>  [7,] 0.6527915 0.3488623 0.7205412
#>  [8,] 0.6354416 2.3861331 0.3293560
#>  [9,] 0.4523859 0.4523859 0.1206830
#> [10,] 0.6702905 1.8229091 3.4128022
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>             [,1]       [,2]        [,3]
#>  [1,] 1.03096051 4.69490247 2.259811723
#>  [2,] 0.05207975 0.01001576 0.052079752
#>  [3,] 4.32495409 0.02270294 2.077200862
#>  [4,] 0.56915168 0.71250130 0.091648747
#>  [5,] 0.30268853 0.52615632 1.865414302
#>  [6,] 2.71950297 3.15191277 0.435502182
#>  [7,] 0.01838536 0.01838536 0.018385357
#>  [8,] 0.25148441 2.08939566 7.077296963
#>  [9,] 1.36651738 1.68626517 0.999457645
#> [10,] 0.50746434 0.50746434 0.006677573
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.819857836 0.22799615 0.05473202
#>  [2,] 0.177847896 0.05913003 0.05265386
#>  [3,] 0.595879578 0.24188855 2.23972589
#>  [4,] 2.195160719 3.02249762 2.05715968
#>  [5,] 0.083895134 0.08389513 0.08389513
#>  [6,] 0.006522685 1.26642762 0.44611466
#>  [7,] 3.146747762 3.43654150 1.00242190
#>  [8,] 1.251201418 1.25120142 0.21841483
#>  [9,] 0.547730592 0.54773059 0.47187438
#> [10,] 4.539101306 4.53910131 0.02054713
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.3898697 0.7423775 1.5030574
#>  [2,] 0.6074422 0.6074422 0.4070949
#>  [3,] 2.8322854 0.9994443 4.2414863
#>  [4,] 0.2312720 0.2312720 0.2312720
#>  [5,] 1.3371237 0.4330089 1.3371237
#>  [6,] 4.1925799 1.7456272 2.1059931
#>  [7,] 1.2882189 1.2822519 0.2123318
#>  [8,] 2.2171473 0.4709548 0.7718791
#>  [9,] 0.1963576 0.5991021 1.7236590
#> [10,] 0.4395958 0.4395958 0.4395958

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43878604 1.43878604 0.95841872
#>  [2,] 0.15875629 0.15875629 0.15875629
#>  [3,] 0.30824564 0.30824564 0.30824564
#>  [4,] 0.08946117 0.08946117 0.08946117
#>  [5,] 0.09727175 0.43900362 0.09727175
#>  [6,] 0.13374030 0.13374030 0.13374030
#>  [7,] 0.55371025 0.05774378 0.93516484
#>  [8,] 0.76627756 0.76627756 0.54306363
#>  [9,] 1.74503204 1.74503204 1.74503204
#> [10,] 0.21984501 0.17757443 0.03631441
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.18022838 0.18022838 0.180228384
#>  [2,] 0.27816629 0.27216545 0.278166288
#>  [3,] 0.73572812 0.73572812 0.735728118
#>  [4,] 0.04294923 0.04294923 0.042949230
#>  [5,] 1.03075614 1.03075614 1.030756137
#>  [6,] 0.54555817 0.38001259 0.545558166
#>  [7,] 2.31022896 1.14002844 1.661767889
#>  [8,] 0.11898906 0.11898906 1.534763047
#>  [9,] 0.41407661 0.92347151 0.002021954
#> [10,] 0.33210476 0.33210476 0.332104761
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.602384831 0.602384831 0.310267000
#>  [2,] 0.773278678 0.196000882 0.773278678
#>  [3,] 0.003083572 0.003083572 0.003083572
#>  [4,] 0.494636456 0.494636456 0.630730703
#>  [5,] 0.766660912 0.766660912 0.649857909
#>  [6,] 0.087457247 0.087457247 0.087457247
#>  [7,] 0.763712825 1.183964140 1.183964140
#>  [8,] 0.268526502 0.615541943 0.615541943
#>  [9,] 2.005690878 0.082321567 0.082321567
#> [10,] 0.170711654 0.170711654 0.144312590
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.195095393 0.07546181 0.19509539
#>  [2,] 1.437181836 1.43718184 1.11551110
#>  [3,] 0.230493022 0.09238033 0.18658469
#>  [4,] 0.079050291 0.07905029 0.07905029
#>  [5,] 0.315205102 0.29598767 0.66268608
#>  [6,] 0.366228887 0.35643624 0.40389574
#>  [7,] 0.007343671 0.08535571 0.08535571
#>  [8,] 0.682898513 1.74178361 0.67055869
#>  [9,] 0.298075444 0.29807544 0.11114714
#> [10,] 1.223607598 1.22360760 0.29202662
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.007232571 0.36748432 0.36748432
#>  [2,] 2.255974173 1.72230874 0.14191818
#>  [3,] 0.732179990 1.14320233 0.11515648
#>  [4,] 0.059037205 0.88796949 0.05903720
#>  [5,] 0.707139423 0.15299512 0.83622356
#>  [6,] 0.181332038 0.18133204 0.18133204
#>  [7,] 0.157180000 0.15718000 0.05152888
#>  [8,] 0.557344987 0.55734499 0.55734499
#>  [9,] 0.259312735 0.04760846 0.25931273
#> [10,] 0.053074590 0.98198628 0.06609653

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33726452 0.33726452 0.33726452
#>  [2,] 0.66535957 0.61618590 0.66535957
#>  [3,] 0.10172757 0.12088296 0.12088296
#>  [4,] 0.19903591 0.19903591 0.19903591
#>  [5,] 0.04601910 0.04601910 0.04601910
#>  [6,] 0.06575302 0.06575302 0.06575302
#>  [7,] 1.05119623 1.44468506 0.19096256
#>  [8,] 0.62912896 0.84184723 1.04367945
#>  [9,] 0.88507838 0.43947639 0.43947639
#> [10,] 0.52687633 0.27257838 0.08893096
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.21440797 0.21440797 0.214407969
#>  [2,] 0.04717123 0.06069733 0.047171228
#>  [3,] 0.05509800 0.01873543 0.055098003
#>  [4,] 0.14588497 1.50238425 0.322116012
#>  [5,] 0.31848271 0.31848271 0.077733895
#>  [6,] 0.21697355 0.43198244 0.004387559
#>  [7,] 0.61038282 0.23968534 0.809928912
#>  [8,] 1.02277423 0.96063242 0.757920498
#>  [9,] 1.70533629 0.20160569 0.623067556
#> [10,] 0.16634289 0.16634289 0.166342887
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.88198751 0.89163438 0.17994394
#>  [2,] 0.15399980 0.15399980 0.15399980
#>  [3,] 0.06678615 0.06678615 0.35442486
#>  [4,] 1.56412475 0.16559948 0.18560420
#>  [5,] 0.02396977 0.02396977 0.74870038
#>  [6,] 0.04858936 0.04858936 0.04858936
#>  [7,] 0.15492229 0.02228997 0.15492229
#>  [8,] 0.10615075 0.31603467 0.10615075
#>  [9,] 0.27433234 0.27433234 0.02702115
#> [10,] 0.26245370 0.26245370 0.26245370
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.53616033 0.20304218 0.38039810
#>  [2,] 0.09754615 0.09754615 0.09754615
#>  [3,] 1.59770640 0.23140892 0.23140892
#>  [4,] 0.18069414 1.79891979 0.18069414
#>  [5,] 0.15331786 0.15331786 0.03164469
#>  [6,] 0.39499357 0.39499357 0.39499357
#>  [7,] 1.01648584 0.40119478 1.01648584
#>  [8,] 0.64936963 0.21203381 1.31529467
#>  [9,] 0.53252041 0.82785829 0.68402803
#> [10,] 0.03677058 0.08833271 0.37317493

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60439628 0.24534799 0.08290730
#>  [2,] 0.01365481 0.52271377 0.52271377
#>  [3,] 0.01682503 0.06938391 0.34286010
#>  [4,] 0.13107297 0.13107297 0.08521192
#>  [5,] 0.12790370 1.17492731 0.04689627
#>  [6,] 0.25582129 0.61716953 0.25582129
#>  [7,] 0.44767231 0.17129254 0.44767231
#>  [8,] 0.28151253 0.28151253 0.68734447
#>  [9,] 0.37054766 0.12096478 0.57059013
#> [10,] 0.10702999 0.32327819 0.14560256
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>               [,1]      [,2]         [,3]
#>  [1,] 0.3582564297 0.3582564 0.2555993019
#>  [2,] 0.1230855322 1.2947873 0.3595778904
#>  [3,] 0.1981089623 0.3405008 0.1981089623
#>  [4,] 0.1571382155 1.3331779 0.3483195681
#>  [5,] 0.9324147228 0.8501088 0.5014288345
#>  [6,] 0.0008268023 0.5184790 0.0008268023
#>  [7,] 0.4610640322 0.3160334 0.3160334260
#>  [8,] 0.3304014480 0.3330632 0.3304014480
#>  [9,] 0.4436651943 0.4436652 0.1630511208
#> [10,] 0.3186897331 0.1626951 0.5819494962
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.256052261 0.316339632 0.009093025
#>  [2,] 0.010048691 1.110571078 0.507627797
#>  [3,] 0.355381751 0.382703180 0.122509087
#>  [4,] 0.544190751 0.544190751 0.456667847
#>  [5,] 0.202176330 0.084123787 0.084123787
#>  [6,] 0.038523301 0.038523301 0.025560366
#>  [7,] 0.296944422 0.237586622 0.042159557
#>  [8,] 0.009333539 0.009333539 0.009333539
#>  [9,] 0.072798900 0.090170102 0.455834834
#> [10,] 0.171290091 0.171290091 0.171290091
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.124559254 0.7314686 0.1254696
#>  [2,] 0.333473600 0.9724513 0.2940948
#>  [3,] 0.002090452 0.9910412 1.4634919
#>  [4,] 1.082805801 0.3467133 0.1912670
#>  [5,] 0.347806966 0.4591735 0.3478070
#>  [6,] 0.512981295 0.4403071 0.5796476
#>  [7,] 0.440702581 0.3444184 0.1538667
#>  [8,] 0.204133642 0.3677486 0.2448675
#>  [9,] 0.922011404 0.5324610 0.5740413
#> [10,] 0.377107871 0.3771079 0.3771079

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1084803 0.69328244 1.42806769
#>  [2,] 0.1206997 0.12069968 0.12069968
#>  [3,] 0.2582257 0.30946497 0.25011793
#>  [4,] 0.1958664 0.62512582 0.21677763
#>  [5,] 0.1452670 0.04748299 0.04748299
#>  [6,] 0.2930755 0.42794406 0.64266422
#>  [7,] 0.1139010 0.11390099 0.11390099
#>  [8,] 0.5162039 0.29068826 0.51620393
#>  [9,] 0.4871739 0.96559924 0.02942080
#> [10,] 0.2271054 0.47212244 0.09565434
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]         [,3]
#>  [1,] 0.06878362 0.41417283 0.0006254596
#>  [2,] 0.11299651 0.11299651 0.1129965051
#>  [3,] 1.00172373 0.29261051 0.2926105079
#>  [4,] 1.08682497 0.14076546 0.1407654589
#>  [5,] 0.87987301 0.36921739 0.4980262984
#>  [6,] 1.14741847 0.72571064 0.7257106419
#>  [7,] 0.75840281 0.28403297 0.1869021000
#>  [8,] 0.42602950 0.02444707 0.4260295041
#>  [9,] 0.03270832 0.55564861 0.2570362859
#> [10,] 0.41985065 0.36591280 0.4730044075
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04649921 0.14092547 0.10385590
#>  [2,] 0.20338195 0.98416847 0.41905596
#>  [3,] 0.40099383 0.36746768 0.05642138
#>  [4,] 0.07981522 0.11867669 0.11867669
#>  [5,] 0.15852247 0.15852247 0.15852247
#>  [6,] 0.34706407 0.34706407 0.34706407
#>  [7,] 0.06416503 0.06416503 0.06416503
#>  [8,] 0.16373250 0.17487762 0.83494503
#>  [9,] 0.79845186 0.55229643 0.29083772
#> [10,] 0.68136374 0.70503620 0.67665072
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4310560 0.07680369 0.1382074
#>  [2,] 0.2407039 0.24070389 0.3852057
#>  [3,] 0.5053546 0.50535456 0.5053546
#>  [4,] 0.0621858 0.06218580 0.0621858
#>  [5,] 0.2928433 0.29284326 0.2928433
#>  [6,] 0.1448094 0.48166991 0.4816699
#>  [7,] 1.2768402 0.18042954 0.8348768
#>  [8,] 0.8435874 0.77206453 0.8059091
#>  [9,] 0.2071777 0.04615575 0.2071777
#> [10,] 0.4631150 0.20138151 0.2840051