Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4373953 0.43634264 2.8827220
#>  [2,] 0.7886801 0.06494495 0.8768107
#>  [3,] 2.2830662 0.75714275 1.6898054
#>  [4,] 1.0934866 2.89623822 1.2088242
#>  [5,] 1.1970292 1.28369751 1.2767744
#>  [6,] 0.3914598 0.39145980 0.3914598
#>  [7,] 0.7174553 0.79687743 5.0579407
#>  [8,] 0.9680633 4.80594828 6.0274824
#>  [9,] 0.4817761 0.48177609 0.4817761
#> [10,] 3.1320207 2.54930960 0.3038483
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2803102 0.2803102 0.2803102
#>  [2,] 4.5010979 4.5010979 4.5010979
#>  [3,] 0.4520403 0.4520403 0.4520403
#>  [4,] 3.7455505 3.7455505 3.7455505
#>  [5,] 0.1991404 0.1991404 0.1991404
#>  [6,] 0.2695482 0.2695482 0.2695482
#>  [7,] 0.7815419 0.7815419 0.7815419
#>  [8,] 0.9294773 0.9294773 0.9294773
#>  [9,] 2.2047118 2.2047118 2.2047118
#> [10,] 0.8273697 0.8273697 0.8273697
## independence
rpextmo(10, 3, b = 1)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1795831 0.22093092 1.08190456
#>  [2,] 2.2000192 2.23335381 0.23639187
#>  [3,] 0.5336435 1.07603957 0.55566849
#>  [4,] 0.5359810 0.21684626 2.62419375
#>  [5,] 0.8905420 0.32844877 1.50914184
#>  [6,] 1.2486922 0.78226540 0.42970786
#>  [7,] 2.8854968 2.19402078 0.03954590
#>  [8,] 0.7227703 0.04025167 0.08934082
#>  [9,] 0.8981007 0.53419839 0.15713823
#> [10,] 0.7902587 1.27510383 1.60659183

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]        [,3]
#>  [1,] 6.3325043 1.3768574 1.470712071
#>  [2,] 0.5655269 0.4572926 0.707634185
#>  [3,] 1.9884082 1.9884082 0.004454634
#>  [4,] 2.7566483 0.6360820 1.224419564
#>  [5,] 1.9171514 5.2224985 0.836962795
#>  [6,] 0.3646553 0.7950195 1.632858975
#>  [7,] 3.9369617 2.0135494 3.892883077
#>  [8,] 0.7905379 0.4728175 0.198267023
#>  [9,] 1.2982696 0.5811209 4.709297127
#> [10,] 0.7003725 0.4828736 0.014589166
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05321247 0.05321247 0.05321247
#>  [2,] 1.59340739 1.59340739 1.59340739
#>  [3,] 5.45018540 5.45018540 5.45018540
#>  [4,] 0.78916841 0.78916841 0.78916841
#>  [5,] 0.03668500 0.03668500 0.03668500
#>  [6,] 0.09857632 0.09857632 0.09857632
#>  [7,] 1.16443807 1.16443807 1.16443807
#>  [8,] 0.04342502 0.04342502 0.04342502
#>  [9,] 0.01512214 0.01512214 0.01512214
#> [10,] 0.70607432 0.70607432 0.70607432
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1511894 0.66817698 1.55395359
#>  [2,] 2.3350909 0.45493659 0.07097353
#>  [3,] 1.7390445 0.71071927 2.66096260
#>  [4,] 1.4771793 1.98266230 0.36963936
#>  [5,] 0.4496496 2.28116984 1.11353495
#>  [6,] 0.3563122 0.02310282 0.14004208
#>  [7,] 0.5894208 3.63072433 0.57552953
#>  [8,] 1.1007906 2.51049558 0.16744960
#>  [9,] 1.2204797 1.00940853 0.75501311
#> [10,] 1.5977861 0.05445837 0.91392961

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 2.6675029 3.44083557 1.7394676
#>  [2,] 2.0108741 4.17462691 4.1746269
#>  [3,] 0.5253858 1.86469762 1.9087416
#>  [4,] 2.4046445 1.31047598 2.4046445
#>  [5,] 2.2583242 0.92701671 0.8163850
#>  [6,] 1.4958069 0.89882283 1.4958069
#>  [7,] 1.5280611 1.52806110 0.9925734
#>  [8,] 3.0045662 0.17608269 1.8555876
#>  [9,] 0.6034724 0.05967433 0.1675905
#> [10,] 1.7238311 6.97059952 1.2627854
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4521901 0.4521901 0.4521901
#>  [2,] 1.4507766 1.4507766 1.4507766
#>  [3,] 0.3707842 0.3707842 0.3707842
#>  [4,] 1.7019948 1.7019948 1.7019948
#>  [5,] 3.8282716 3.8282716 3.8282716
#>  [6,] 0.2220400 0.2220400 0.2220400
#>  [7,] 1.8710768 1.8710768 1.8710768
#>  [8,] 0.1905101 0.1905101 0.1905101
#>  [9,] 0.5606260 0.5606260 0.5606260
#> [10,] 0.3720590 0.3720590 0.3720590
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7187991 0.04502887 0.3466566
#>  [2,] 0.9725578 0.14249219 1.6505741
#>  [3,] 0.2691931 1.01735018 1.7849753
#>  [4,] 1.1236034 0.84117896 2.0083391
#>  [5,] 0.3390136 0.91793636 0.4783296
#>  [6,] 0.5739502 0.26767576 0.0593693
#>  [7,] 1.0342439 0.33917304 0.1213083
#>  [8,] 1.1565385 1.02942890 0.9500125
#>  [9,] 1.4764833 0.46434797 0.2466086
#> [10,] 0.5944723 0.06148563 0.6471353

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27164974 0.27164974 0.27164974
#>  [2,] 1.18203499 4.09896497 1.15200555
#>  [3,] 1.78086032 1.53644635 1.95512275
#>  [4,] 0.04334944 0.04334944 0.04334944
#>  [5,] 3.35005169 0.27289086 0.35202005
#>  [6,] 0.77526092 1.94151169 2.31081152
#>  [7,] 0.08916562 0.08916562 0.08916562
#>  [8,] 0.08228779 0.08228779 0.08228779
#>  [9,] 0.35671308 0.35671308 0.35671308
#> [10,] 1.81698872 1.81698872 0.96907343
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.71091065 1.71091065 1.71091065
#>  [2,] 1.26716765 1.26716765 1.26716765
#>  [3,] 0.76715409 0.76715409 0.76715409
#>  [4,] 0.17313228 0.17313228 0.17313228
#>  [5,] 1.09275885 1.09275885 1.09275885
#>  [6,] 0.05127626 0.05127626 0.05127626
#>  [7,] 2.42322335 2.42322335 2.42322335
#>  [8,] 0.62906928 0.62906928 0.62906928
#>  [9,] 0.95321433 0.95321433 0.95321433
#> [10,] 2.17422843 2.17422843 2.17422843
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.39524615 2.37788579 2.75889417
#>  [2,] 0.80401188 0.06821121 1.15507406
#>  [3,] 0.25344979 0.57286333 0.45138342
#>  [4,] 0.44959430 2.32467478 1.43747697
#>  [5,] 3.38429598 1.38499692 0.20868282
#>  [6,] 1.80827034 0.96673729 1.26906877
#>  [7,] 0.05124852 0.27283589 0.13698192
#>  [8,] 0.33665020 4.00480625 0.04492185
#>  [9,] 0.17532088 3.09567805 0.14188258
#> [10,] 0.38645518 2.44195082 0.50353056

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1641863 1.7354163 0.1000747
#>  [2,] 0.2698279 1.0999987 0.5005552
#>  [3,] 0.9399966 3.0914922 1.5540080
#>  [4,] 2.0997966 2.1602628 2.1602628
#>  [5,] 1.0854804 0.1407112 1.2891768
#>  [6,] 0.2177306 0.1154485 2.1294112
#>  [7,] 0.5365224 1.8703726 2.3749126
#>  [8,] 0.6724244 0.6724244 0.6724244
#>  [9,] 0.9158161 2.4321732 1.2168480
#> [10,] 0.7612453 0.7612453 0.7612453
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.24297177 2.24297177 2.24297177
#>  [2,] 0.25545802 0.25545802 0.25545802
#>  [3,] 0.83325737 0.83325737 0.83325737
#>  [4,] 0.01397109 0.01397109 0.01397109
#>  [5,] 2.64314832 2.64314832 2.64314832
#>  [6,] 0.40754171 0.40754171 0.40754171
#>  [7,] 0.27103565 0.27103565 0.27103565
#>  [8,] 0.93646578 0.93646578 0.93646578
#>  [9,] 0.63732351 0.63732351 0.63732351
#> [10,] 1.29755385 1.29755385 1.29755385
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]      [,2]         [,3]
#>  [1,] 1.27893382 1.9988361 0.6411412613
#>  [2,] 0.33701967 0.4686343 0.2031476999
#>  [3,] 2.71065990 0.6794988 2.5021651641
#>  [4,] 1.37005959 0.1643941 0.0006639151
#>  [5,] 1.15611110 0.7065841 1.1675480329
#>  [6,] 0.62001177 0.1830385 0.3825132628
#>  [7,] 0.92367658 0.9878650 2.0995930689
#>  [8,] 0.05434405 0.0524316 0.2115319314
#>  [9,] 0.25575906 1.4806448 0.8845768379
#> [10,] 2.01931698 0.4343069 0.6680384536

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01729894 0.21651959 1.93321295
#>  [2,] 0.57756325 0.01286481 0.57756325
#>  [3,] 1.76239585 0.03251210 0.17706307
#>  [4,] 0.80778266 0.17238557 0.84926489
#>  [5,] 1.01422597 0.65415780 1.82225074
#>  [6,] 5.57418892 2.76094755 1.54461824
#>  [7,] 1.24963047 1.32627111 4.49544169
#>  [8,] 0.16799424 0.16799424 0.16799424
#>  [9,] 0.45179688 0.73725635 0.45251743
#> [10,] 1.14951736 4.70835281 0.09408686
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4235670 0.4235670 0.4235670
#>  [2,] 0.7731452 0.7731452 0.7731452
#>  [3,] 0.2739900 0.2739900 0.2739900
#>  [4,] 0.4750636 0.4750636 0.4750636
#>  [5,] 0.3358324 0.3358324 0.3358324
#>  [6,] 1.9430134 1.9430134 1.9430134
#>  [7,] 0.3215183 0.3215183 0.3215183
#>  [8,] 0.9270247 0.9270247 0.9270247
#>  [9,] 0.9569965 0.9569965 0.9569965
#> [10,] 0.3769991 0.3769991 0.3769991
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.20067592 0.9279579 2.5422680
#>  [2,] 0.98780531 0.3865853 0.2527112
#>  [3,] 0.48174950 1.3186072 1.9665465
#>  [4,] 0.07525228 0.4548460 2.9121436
#>  [5,] 0.47564090 0.5002131 0.3306468
#>  [6,] 1.80531927 0.7598651 0.1256301
#>  [7,] 1.02052062 1.1374115 0.8244696
#>  [8,] 0.72124874 0.2313291 0.4701654
#>  [9,] 2.21218672 3.7098517 0.8476823
#> [10,] 0.87917869 0.9028296 0.1124584

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.71934641 1.92581329 1.80236760
#>  [2,] 0.47000342 0.12296434 1.03485760
#>  [3,] 1.89953251 1.89953251 1.15049445
#>  [4,] 1.88763317 0.19309870 0.71726494
#>  [5,] 0.30231745 0.27754040 0.60928050
#>  [6,] 0.09984623 0.09984623 0.09984623
#>  [7,] 0.38749662 0.38749662 0.03049303
#>  [8,] 0.40433649 0.16648897 0.15379195
#>  [9,] 0.83668278 1.30546723 0.03997732
#> [10,] 1.56532418 2.97880213 2.97880213
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1316198 1.1316198 1.1316198
#>  [2,] 0.1741270 0.1741270 0.1741270
#>  [3,] 1.2135518 1.2135518 1.2135518
#>  [4,] 0.5764774 0.5764774 0.5764774
#>  [5,] 1.2581101 1.2581101 1.2581101
#>  [6,] 0.3160529 0.3160529 0.3160529
#>  [7,] 0.1549415 0.1549415 0.1549415
#>  [8,] 0.2872751 0.2872751 0.2872751
#>  [9,] 3.8882113 3.8882113 3.8882113
#> [10,] 0.8752314 0.8752314 0.8752314
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.221001563 0.08054570 0.33484126
#>  [2,] 0.765295516 0.16134678 0.18983578
#>  [3,] 2.505768312 0.05257216 1.12009349
#>  [4,] 1.320554687 0.10549036 0.92467278
#>  [5,] 0.643149437 0.34050685 0.96619504
#>  [6,] 0.006270506 0.92717204 0.07079563
#>  [7,] 2.488998097 1.15761413 0.36612817
#>  [8,] 6.840301020 0.60367349 0.01812083
#>  [9,] 0.429965494 3.45810996 0.29886254
#> [10,] 0.098854735 0.04098939 0.03269726

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 3.43637270 0.1398972 1.26078377
#>  [2,] 0.20290962 0.2764548 0.16737130
#>  [3,] 0.06522311 1.8435874 0.60207656
#>  [4,] 4.28144582 1.3559671 4.23332815
#>  [5,] 0.61842465 0.9115367 0.91153668
#>  [6,] 3.23986771 3.2442828 1.54732450
#>  [7,] 0.34768671 0.1602408 0.60408264
#>  [8,] 1.55735777 1.8199802 1.31848650
#>  [9,] 3.17361046 5.7448627 0.07142468
#> [10,] 2.54582911 3.1708884 0.12431528
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9920473 0.9920473 0.9920473
#>  [2,] 1.0359146 1.0359146 1.0359146
#>  [3,] 0.3792490 0.3792490 0.3792490
#>  [4,] 0.1799308 0.1799308 0.1799308
#>  [5,] 2.1988983 2.1988983 2.1988983
#>  [6,] 3.8249735 3.8249735 3.8249735
#>  [7,] 0.9344195 0.9344195 0.9344195
#>  [8,] 0.6524438 0.6524438 0.6524438
#>  [9,] 2.7199078 2.7199078 2.7199078
#> [10,] 1.0216890 1.0216890 1.0216890
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3274434 0.2088192 0.6190528
#>  [2,] 2.3054642 0.4937882 0.2622881
#>  [3,] 1.1273682 0.4407213 0.3569157
#>  [4,] 1.2327326 0.2186863 0.5118617
#>  [5,] 3.6696217 0.1350140 0.3136188
#>  [6,] 2.1669181 3.2420753 1.8151041
#>  [7,] 0.3327839 0.5413013 0.5804638
#>  [8,] 0.2021845 0.2458703 2.1331240
#>  [9,] 1.8126256 1.9055191 1.1624937
#> [10,] 0.8910700 1.7215572 0.1379760

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3311355 0.3114331 1.4158470
#>  [2,] 0.7580382 0.5087388 0.7580382
#>  [3,] 0.1204060 0.3788364 2.0654129
#>  [4,] 0.2493854 0.1189950 0.2493854
#>  [5,] 0.8500644 0.8500644 0.4556580
#>  [6,] 0.2186071 0.4245134 0.4245134
#>  [7,] 0.2747198 0.3585737 2.1006104
#>  [8,] 1.0967654 1.8470239 0.4388212
#>  [9,] 0.5119913 0.5203314 2.0712922
#> [10,] 1.6769122 1.2819776 0.5219541
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.10472439 1.10472439 1.10472439
#>  [2,] 0.61310124 0.61310124 0.61310124
#>  [3,] 0.26511516 0.26511516 0.26511516
#>  [4,] 0.28483232 0.28483232 0.28483232
#>  [5,] 0.10045854 0.10045854 0.10045854
#>  [6,] 1.12497327 1.12497327 1.12497327
#>  [7,] 0.01838942 0.01838942 0.01838942
#>  [8,] 0.73430267 0.73430267 0.73430267
#>  [9,] 1.59551937 1.59551937 1.59551937
#> [10,] 2.39027227 2.39027227 2.39027227
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.09481978 2.00560069 0.8837483
#>  [2,] 0.26243011 1.12721218 0.4367280
#>  [3,] 0.74741879 0.01312233 0.7829473
#>  [4,] 1.02422440 0.33842708 0.7086409
#>  [5,] 2.91998051 0.37256758 1.3408893
#>  [6,] 0.95448918 0.75556309 1.5338000
#>  [7,] 0.80714143 0.65248477 0.1009925
#>  [8,] 0.75176760 0.10169987 0.8329790
#>  [9,] 0.36043972 1.06975062 0.6241453
#> [10,] 0.06393501 1.67680685 3.0854158

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4375482 0.4375482 0.4375482
#>  [2,] 0.2684121 0.2684121 0.2684121
#>  [3,] 1.3844085 0.5107905 0.1505195
#>  [4,] 0.9038790 3.0911302 0.2733313
#>  [5,] 3.1979572 3.3879776 1.1759666
#>  [6,] 0.6905745 0.6905745 0.3507870
#>  [7,] 0.1234220 0.1234220 0.1234220
#>  [8,] 4.1209617 1.6584097 6.8460577
#>  [9,] 2.4157431 3.1981166 3.1981166
#> [10,] 0.8308700 0.8308700 0.8308700
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2015185 0.2015185 0.2015185
#>  [2,] 2.0688218 2.0688218 2.0688218
#>  [3,] 1.6758822 1.6758822 1.6758822
#>  [4,] 0.7231557 0.7231557 0.7231557
#>  [5,] 0.8818130 0.8818130 0.8818130
#>  [6,] 0.7546504 0.7546504 0.7546504
#>  [7,] 0.9440498 0.9440498 0.9440498
#>  [8,] 1.2182283 1.2182283 1.2182283
#>  [9,] 0.8714029 0.8714029 0.8714029
#> [10,] 0.7915743 0.7915743 0.7915743
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6351182 0.2995906 0.3843116
#>  [2,] 1.1862819 0.3036276 0.5500009
#>  [3,] 0.4783184 2.8776139 0.4256958
#>  [4,] 0.5790113 0.1414120 1.1234189
#>  [5,] 0.1669628 2.5859850 2.7877658
#>  [6,] 0.1108359 0.3611685 1.0851232
#>  [7,] 0.6768940 2.0614989 2.2260708
#>  [8,] 0.4895170 0.4621187 1.1785769
#>  [9,] 0.4335932 0.6409583 0.1132238
#> [10,] 0.6003685 0.4563615 0.4336124

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.35953121 0.6614849 0.8212987
#>  [2,] 0.74738561 0.5581867 0.1423125
#>  [3,] 0.90715765 1.0348443 1.0348443
#>  [4,] 2.71608046 1.0548213 0.4324538
#>  [5,] 0.76764167 0.1292925 0.9437018
#>  [6,] 0.08326145 0.4463677 0.7499683
#>  [7,] 0.28358014 1.2403114 2.1398017
#>  [8,] 0.46280312 0.4628031 0.4628031
#>  [9,] 1.73311217 1.3186184 2.2459814
#> [10,] 0.86060955 0.8724314 0.6023563
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1216947 0.42726396 3.67231866
#>  [2,] 2.0144550 0.08041133 0.71900904
#>  [3,] 0.5220223 0.84394242 1.83806583
#>  [4,] 0.2295662 0.22956616 1.27334621
#>  [5,] 0.7573460 0.80705943 0.80705943
#>  [6,] 0.6499290 0.60024089 0.64992896
#>  [7,] 1.0884652 0.43065601 0.09748444
#>  [8,] 1.0480205 0.79622996 0.98718208
#>  [9,] 1.2537379 0.44770822 1.19720835
#> [10,] 1.5464371 0.59515801 0.82248419
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5055924 0.5395166 0.46190311
#>  [2,] 1.5135282 0.7386892 0.64714474
#>  [3,] 0.5776859 0.2981203 0.81987783
#>  [4,] 0.6862052 0.4402941 0.69534669
#>  [5,] 0.1336893 1.2422372 0.08801338
#>  [6,] 0.1298596 0.1298596 0.12985958
#>  [7,] 0.6652130 0.6652130 0.66521301
#>  [8,] 0.2918901 0.2918901 0.29189008
#>  [9,] 0.3896248 0.3077695 0.86276495
#> [10,] 0.6799621 0.2664218 0.18723960
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5599525 0.5599696 0.03891889
#>  [2,] 0.9524678 0.5016769 0.69908728
#>  [3,] 0.6338574 0.2461838 0.63385737
#>  [4,] 0.3799502 0.9163758 1.44736402
#>  [5,] 1.3792461 0.2008421 2.20667654
#>  [6,] 1.0015022 0.1062272 0.71598257
#>  [7,] 0.9145177 0.3855083 0.75181499
#>  [8,] 1.5460711 0.4188017 0.38937979
#>  [9,] 1.3545788 0.2330845 0.66901056
#> [10,] 0.1875266 0.1875266 1.41411138
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.88933343 0.79579788 0.13313840
#>  [2,] 0.27440435 0.15574295 0.15574295
#>  [3,] 0.12129908 0.12129908 0.12129908
#>  [4,] 2.63409610 0.33268834 0.15914184
#>  [5,] 0.11297629 0.73984427 1.12835997
#>  [6,] 0.87181009 0.78362984 1.49855216
#>  [7,] 1.98597802 2.01190350 0.55461183
#>  [8,] 0.08687296 0.50772771 0.08687296
#>  [9,] 0.03020157 0.03020157 0.03020157
#> [10,] 0.27085756 0.27085756 0.86946867

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27209239 0.21808195 3.37639047
#>  [2,] 2.99244980 0.44284115 0.42301636
#>  [3,] 0.09825348 0.09825348 0.03867773
#>  [4,] 0.78042662 1.61473505 3.00641062
#>  [5,] 0.82391702 1.02947054 5.66267944
#>  [6,] 1.55342198 0.40781917 0.75475183
#>  [7,] 2.49014369 0.18185787 1.63147192
#>  [8,] 2.01772155 2.01772155 0.24239392
#>  [9,] 0.36217149 2.71340342 2.71340342
#> [10,] 1.04487677 0.22193670 0.13954509
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09027781 0.09027781 0.09027781
#>  [2,] 0.79996642 1.32903417 1.32903417
#>  [3,] 0.31258998 0.59807731 0.59807731
#>  [4,] 0.22902075 1.52018057 0.72546379
#>  [5,] 0.01514264 0.01514264 0.01514264
#>  [6,] 0.39728578 2.91856333 0.16191168
#>  [7,] 6.66583468 1.30865823 0.66854613
#>  [8,] 0.01325524 0.30639280 0.62403390
#>  [9,] 4.10829040 4.10829040 0.85069992
#> [10,] 0.42281278 1.70616772 0.53635964
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>            [,1]        [,2]       [,3]
#>  [1,] 0.3562246 2.886380315 1.11928513
#>  [2,] 1.6044182 7.262816726 0.08883505
#>  [3,] 0.2567613 1.777306670 1.49565413
#>  [4,] 1.8011698 1.295741252 1.43174860
#>  [5,] 1.1762513 1.469897318 1.46989732
#>  [6,] 1.0696009 0.002495005 2.29004801
#>  [7,] 0.1055772 0.105577150 0.10557715
#>  [8,] 0.3500966 0.350096622 0.35009662
#>  [9,] 1.0874181 0.157530017 1.08741809
#> [10,] 2.7535953 0.910915362 0.64945567
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0862399 0.05979302 1.70907587
#>  [2,] 1.4696983 2.68278291 1.34777998
#>  [3,] 1.4791345 0.61527723 0.18925102
#>  [4,] 2.0873548 2.18264828 2.18264828
#>  [5,] 0.1430832 0.14308320 0.14308320
#>  [6,] 0.3245106 2.89529907 0.93536038
#>  [7,] 0.8343852 0.41846408 0.57439157
#>  [8,] 1.7279147 2.06772682 0.08420652
#>  [9,] 3.2606743 0.44753101 2.15027560
#> [10,] 0.4340925 0.87380491 2.72578817
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6091823 0.60918229 0.37368363
#>  [2,] 0.4942288 0.09297164 0.49422884
#>  [3,] 3.0276801 1.04510623 2.30276241
#>  [4,] 3.4533902 0.01681137 0.17941650
#>  [5,] 1.1108944 0.36318119 1.11089440
#>  [6,] 0.8735527 0.57561008 1.24289170
#>  [7,] 0.9843364 1.72828240 0.20397846
#>  [8,] 0.7822656 0.69988614 1.73791638
#>  [9,] 0.2551042 0.95240959 3.34949805
#> [10,] 1.4279942 1.42799418 0.05123788

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.993985087 1.513188202 0.754293945
#>  [2,] 0.117564196 0.117564196 0.117564196
#>  [3,] 0.176449518 0.250238996 0.232537415
#>  [4,] 0.128715943 0.128715943 0.128715943
#>  [5,] 0.313338994 1.227798789 1.937638268
#>  [6,] 0.131155469 0.131155469 0.131155469
#>  [7,] 0.395527766 0.068621939 0.395527766
#>  [8,] 0.369682131 0.369682131 0.369682131
#>  [9,] 0.002143112 0.002143112 0.002143112
#> [10,] 0.110312589 0.110312589 0.110312589
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15231482 0.13092394 0.15231482
#>  [2,] 0.42084683 0.42084683 0.41427406
#>  [3,] 0.56171897 0.12144248 0.56171897
#>  [4,] 0.22787680 0.10693476 0.10693476
#>  [5,] 0.46986300 0.46986300 0.46986300
#>  [6,] 0.11988471 0.11988471 0.11988471
#>  [7,] 0.68933167 0.23462990 0.12997925
#>  [8,] 1.09914647 0.91995359 0.55630265
#>  [9,] 0.05718307 0.05718307 0.05718307
#> [10,] 0.17938319 0.29556747 0.17938319
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.93053364 0.41891120 0.41891120
#>  [2,] 0.76152247 0.21858103 0.76152247
#>  [3,] 0.42630966 0.42630966 0.42630966
#>  [4,] 0.04438869 2.79038581 0.04438869
#>  [5,] 0.02626955 0.02626955 0.01993764
#>  [6,] 0.19861752 0.09682469 0.16655029
#>  [7,] 0.23879074 0.23879074 0.23879074
#>  [8,] 0.03456857 0.03456857 0.03456857
#>  [9,] 0.53751374 0.37759210 0.37759210
#> [10,] 0.11545711 0.46760055 0.35275162
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 1.13035939 1.249237011 0.72440790
#>  [2,] 0.48570566 0.485705659 0.20037096
#>  [3,] 0.08095847 0.005637927 0.08095847
#>  [4,] 0.52220438 0.522204384 0.52220438
#>  [5,] 0.45812073 0.009486656 0.45812073
#>  [6,] 0.28954590 0.547873362 0.54787336
#>  [7,] 1.75304801 1.336958748 0.12517822
#>  [8,] 2.08239899 0.298854717 0.16172610
#>  [9,] 0.13472791 0.579152209 0.24247225
#> [10,] 0.25750554 0.121959189 0.25750554
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37707217 0.91687476 3.96180260
#>  [2,] 0.75327449 0.75327449 0.65761069
#>  [3,] 1.05399271 1.61160611 2.38086334
#>  [4,] 0.11870852 0.92326581 0.55351063
#>  [5,] 0.01839130 0.34997070 0.20460365
#>  [6,] 0.25607053 0.74831617 0.90478631
#>  [7,] 0.75301125 0.75301125 0.10391100
#>  [8,] 0.30378182 0.09114151 0.30378182
#>  [9,] 0.08828286 0.08828286 0.08828286
#> [10,] 0.10476935 0.10476935 0.10476935

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21890898 0.18829081 0.71837763
#>  [2,] 0.13288622 0.10983174 1.38972202
#>  [3,] 0.04439861 0.66922552 0.49276915
#>  [4,] 0.60406371 1.30476416 0.33273863
#>  [5,] 0.17651892 0.17651892 0.25896767
#>  [6,] 0.74884607 0.05828092 1.30153197
#>  [7,] 0.34901759 0.34901759 0.34901759
#>  [8,] 0.01917558 0.01917558 0.01917558
#>  [9,] 0.53754042 0.44165987 0.26270315
#> [10,] 0.15437343 1.90052469 1.90052469
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11987086 0.11987086 0.20732981
#>  [2,] 0.11424368 0.05953170 0.25674240
#>  [3,] 0.15555280 0.06160422 0.47546609
#>  [4,] 0.01845175 0.11316780 0.08938329
#>  [5,] 0.39417614 0.59987440 0.21904707
#>  [6,] 0.01246184 0.28304128 0.28304128
#>  [7,] 0.82189561 0.11318044 0.29415841
#>  [8,] 0.09046277 0.31851580 0.21120368
#>  [9,] 0.34294911 0.06988974 0.06787538
#> [10,] 0.16981684 0.02322944 0.03674876
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.070678322 0.846788503 0.644732681
#>  [2,] 0.403554629 0.413087379 0.572199806
#>  [3,] 0.156723229 0.055058283 0.156723229
#>  [4,] 0.003308863 0.003308863 0.003308863
#>  [5,] 0.277919920 0.277919920 0.270435621
#>  [6,] 0.177792868 0.177792868 0.177792868
#>  [7,] 0.200328598 0.200328598 0.134755233
#>  [8,] 0.289118946 0.127646076 0.289118946
#>  [9,] 0.290949605 0.506176103 0.506176103
#> [10,] 0.534187435 0.534187435 0.229513729
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31030533 0.63632391 1.42263015
#>  [2,] 0.08720137 0.53557300 0.34351285
#>  [3,] 0.84272907 0.04911163 0.04911163
#>  [4,] 0.07359340 0.07359340 0.07359340
#>  [5,] 0.46139838 0.21606785 0.21606785
#>  [6,] 0.38080853 0.56041188 0.56041188
#>  [7,] 0.11393022 0.11393022 0.11393022
#>  [8,] 0.10779467 0.10779467 0.10779467
#>  [9,] 0.85293552 0.91428946 0.56204091
#> [10,] 1.54090996 1.44744007 1.54090996

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>               [,1]      [,2]       [,3]
#>  [1,] 0.2547618986 0.1721221 0.57890963
#>  [2,] 0.0003989352 0.7030412 0.30128702
#>  [3,] 0.9084917141 0.7643360 0.36804774
#>  [4,] 0.8331626372 1.4230327 1.00377487
#>  [5,] 0.1365309501 0.1365310 0.66681809
#>  [6,] 0.7396906186 0.2754086 0.47119149
#>  [7,] 0.6401518823 0.4392928 0.74052485
#>  [8,] 0.8621492904 0.2074662 0.07612915
#>  [9,] 0.6360304908 0.3040102 0.12201102
#> [10,] 0.3011433423 0.1666339 0.20941021
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2206095 0.22060954 0.05432243
#>  [2,] 0.2421909 0.24219090 0.28935701
#>  [3,] 0.4603795 0.14871984 0.46037950
#>  [4,] 0.7918006 0.79180063 0.43368350
#>  [5,] 0.2395487 1.66448142 0.99503022
#>  [6,] 0.6730476 0.09878639 0.09878639
#>  [7,] 0.3983499 0.48950546 0.49256458
#>  [8,] 0.1593447 0.22075319 0.40902297
#>  [9,] 0.2450119 0.02716004 1.05763849
#> [10,] 0.9960809 0.13347719 0.13347719
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.08034807 0.067775937 0.06455608
#>  [2,] 0.18670247 0.006908188 0.50625877
#>  [3,] 0.45746878 0.904144503 0.05425192
#>  [4,] 0.32955026 0.437914119 2.22635054
#>  [5,] 0.15456724 0.154567240 0.15456724
#>  [6,] 0.70786303 0.470210541 0.52838708
#>  [7,] 0.61852989 0.064243904 0.80049207
#>  [8,] 0.03933058 0.039330580 0.37729978
#>  [9,] 0.19024298 0.190242984 0.19024298
#> [10,] 0.56630765 0.566307649 0.56630765
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15430626 0.20948807 0.20948807
#>  [2,] 0.35252950 0.05890512 0.02086937
#>  [3,] 1.19108273 0.11444245 1.57280977
#>  [4,] 0.55333371 0.55333371 0.29084543
#>  [5,] 0.31089481 0.58618427 0.03594009
#>  [6,] 0.06270966 0.06270966 0.06270966
#>  [7,] 0.97442078 0.23966942 0.53668292
#>  [8,] 0.58885096 0.42884311 0.58885096
#>  [9,] 0.26436200 1.18241296 0.71446303
#> [10,] 0.22598923 0.79964423 0.27759390

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.009791511 0.667703392 1.45299964
#>  [2,] 0.569008764 0.569008764 0.86264910
#>  [3,] 0.375238491 0.752132476 0.36358851
#>  [4,] 0.237705872 0.109539248 0.38521424
#>  [5,] 0.049545063 0.085004990 0.04954506
#>  [6,] 0.074306126 0.073309552 0.43591381
#>  [7,] 0.680197535 0.003654024 0.68019754
#>  [8,] 0.383732640 0.314380009 0.93335658
#>  [9,] 0.608900728 0.071762130 0.52922068
#> [10,] 0.021319563 0.183020944 0.09798284
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.27185892 0.062110905 0.22773029
#>  [2,] 0.10360076 0.002063134 0.26033879
#>  [3,] 0.50236638 0.243123791 0.46937420
#>  [4,] 0.79910081 0.275777260 0.29196821
#>  [5,] 0.30025872 0.064996703 0.76591794
#>  [6,] 0.02151247 0.021512466 0.02151247
#>  [7,] 0.31163333 0.311633332 0.11023841
#>  [8,] 0.06787516 0.067875157 0.78089760
#>  [9,] 0.10912388 0.227930339 0.16298406
#> [10,] 0.18715822 0.179789338 0.13705286
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.772074095 0.772074095 0.48914378
#>  [2,] 0.068731644 0.068731644 0.06873164
#>  [3,] 0.007827551 0.007827551 1.40417100
#>  [4,] 0.150139106 0.195561332 0.18757002
#>  [5,] 0.115527046 0.143175216 0.23910614
#>  [6,] 0.147838155 0.147838155 0.14783815
#>  [7,] 0.029683389 0.029683389 0.02968339
#>  [8,] 0.687385534 0.248206270 0.72569539
#>  [9,] 0.147498071 0.245179015 0.01484625
#> [10,] 0.098651787 0.098651787 0.05650343
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.92387586 0.92387586 0.01658742
#>  [2,] 0.25147939 0.05338875 0.21951113
#>  [3,] 0.06913472 0.26385679 0.25803095
#>  [4,] 0.12802876 0.18582425 0.18582425
#>  [5,] 0.76722171 0.59556397 0.59556397
#>  [6,] 0.07634683 0.07634683 0.07634683
#>  [7,] 0.04543809 0.43154115 0.25719618
#>  [8,] 0.47224878 0.13144108 0.47224878
#>  [9,] 0.23379342 0.90169944 0.35911057
#> [10,] 0.10992272 0.10992272 0.10992272