Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > \beta t \}} + \alpha t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>              [,1]      [,2]       [,3]
#>  [1,] 4.511418681 1.1406014 0.22293089
#>  [2,] 1.390776123 0.3240798 3.89470988
#>  [3,] 0.167982562 0.3051310 2.16973078
#>  [4,] 0.264523314 1.9116408 0.12336955
#>  [5,] 0.417601906 0.4176019 0.41760191
#>  [6,] 2.173036251 3.3087914 2.17221299
#>  [7,] 0.935097693 0.9350977 0.09888933
#>  [8,] 1.983517146 0.1110591 2.62326739
#>  [9,] 1.033249700 0.9644955 1.87450168
#> [10,] 0.005409818 2.6128053 0.17976151
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33420501 0.33420501 0.33420501
#>  [2,] 0.56216041 0.56216041 0.56216041
#>  [3,] 0.81258394 0.81258394 0.81258394
#>  [4,] 0.29818545 0.29818545 0.29818545
#>  [5,] 0.18139848 0.18139848 0.18139848
#>  [6,] 0.12635290 0.12635290 0.12635290
#>  [7,] 0.44073068 0.44073068 0.44073068
#>  [8,] 0.48285294 0.48285294 0.48285294
#>  [9,] 1.26707419 1.26707419 1.26707419
#> [10,] 0.03639986 0.03639986 0.03639986
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]        [,2]       [,3]
#>  [1,] 0.07910286 0.023616959 2.52275908
#>  [2,] 0.01757068 0.215853782 0.11810618
#>  [3,] 0.44521289 0.059488121 1.40488025
#>  [4,] 0.41649918 0.009423681 5.41550248
#>  [5,] 2.20629262 1.330885025 0.08317636
#>  [6,] 0.41914099 0.084888789 2.34776753
#>  [7,] 0.09495000 0.078182291 1.29788797
#>  [8,] 0.92142713 1.359605540 3.85994858
#>  [9,] 0.18768297 0.606258528 0.32141258
#> [10,] 0.13391885 1.014198783 1.00966705

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3112498 0.3112498 0.31124982
#>  [2,] 0.7650643 2.1057641 0.82794010
#>  [3,] 1.6146107 2.0238666 0.05886086
#>  [4,] 0.2127515 0.3729808 1.54813658
#>  [5,] 1.1019027 2.2268516 0.82387865
#>  [6,] 3.0078912 0.4648750 3.21601660
#>  [7,] 1.5629262 0.5327663 4.79024494
#>  [8,] 5.1774410 0.6706932 5.17744100
#>  [9,] 0.9747192 0.5551551 0.97471922
#> [10,] 1.0405715 0.4978547 0.31716502
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 4.4576800 4.4576800 4.4576800
#>  [2,] 1.5944980 1.5944980 1.5944980
#>  [3,] 3.9239779 3.9239779 3.9239779
#>  [4,] 3.7766679 3.7766679 3.7766679
#>  [5,] 2.6253523 2.6253523 2.6253523
#>  [6,] 0.7004484 0.7004484 0.7004484
#>  [7,] 0.2280505 0.2280505 0.2280505
#>  [8,] 0.6603363 0.6603363 0.6603363
#>  [9,] 0.5814125 0.5814125 0.5814125
#> [10,] 1.6559217 1.6559217 1.6559217
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2473531 0.21946760 0.16965810
#>  [2,] 1.6795365 0.27415732 0.86961101
#>  [3,] 1.9544535 0.65101579 2.15444233
#>  [4,] 0.7038132 0.26879216 0.77199996
#>  [5,] 1.0683621 0.66588440 2.81338695
#>  [6,] 0.1235690 1.21253776 0.89798675
#>  [7,] 0.3771055 0.39088512 0.02617014
#>  [8,] 0.2757063 0.68329461 0.21836322
#>  [9,] 0.1244120 0.01169289 3.54179904
#> [10,] 0.8151625 3.96073042 0.59824175

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2606366 0.3138956 0.89820104
#>  [2,] 0.5754899 0.5089483 1.01101390
#>  [3,] 4.1231406 3.7121249 0.36066940
#>  [4,] 1.8291318 2.2461606 2.31537154
#>  [5,] 0.4656598 0.9739794 3.41850209
#>  [6,] 3.4807752 0.9553270 0.07131198
#>  [7,] 1.0283113 3.0204666 0.85295579
#>  [8,] 1.0625573 1.0625573 0.08282331
#>  [9,] 0.8111136 0.8111136 0.09718969
#> [10,] 0.5030771 0.8000102 1.13059606
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.82892476 2.82892476 2.82892476
#>  [2,] 2.08432663 2.08432663 2.08432663
#>  [3,] 0.27032272 0.27032272 0.27032272
#>  [4,] 0.06339732 0.06339732 0.06339732
#>  [5,] 0.94148870 0.94148870 0.94148870
#>  [6,] 0.15345894 0.15345894 0.15345894
#>  [7,] 1.40684740 1.40684740 1.40684740
#>  [8,] 1.31495308 1.31495308 1.31495308
#>  [9,] 3.85952871 3.85952871 3.85952871
#> [10,] 0.17472487 0.17472487 0.17472487
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 3.53681984 4.7999274 0.8954240
#>  [2,] 0.59248447 0.7157134 0.5681300
#>  [3,] 0.98763637 1.1730619 0.2053835
#>  [4,] 0.22742112 0.7035774 0.2659089
#>  [5,] 1.82763341 0.6928550 0.5268615
#>  [6,] 1.33759990 0.8356517 1.6876576
#>  [7,] 0.03500652 0.6273572 2.5156333
#>  [8,] 0.18827727 0.5640413 1.7967901
#>  [9,] 0.76857261 2.3067512 2.2979492
#> [10,] 2.17425729 3.9210662 0.9371438

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3015091 1.5558345 1.7586113
#>  [2,] 3.2078256 1.5999923 2.6870857
#>  [3,] 0.9183703 0.3802127 2.2070145
#>  [4,] 0.7520932 1.8345688 2.4514572
#>  [5,] 1.2107829 2.8806310 2.6276395
#>  [6,] 1.1109162 1.1109162 1.1109162
#>  [7,] 0.6823402 0.4957326 0.2911043
#>  [8,] 3.3278239 4.2175297 0.4397411
#>  [9,] 3.1401080 2.5105229 0.1659086
#> [10,] 2.1554173 1.0941873 0.8564152
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.4529281 2.4529281 2.4529281
#>  [2,] 0.5127502 0.5127502 0.5127502
#>  [3,] 0.8430469 0.8430469 0.8430469
#>  [4,] 0.4968409 0.4968409 0.4968409
#>  [5,] 0.2884727 0.2884727 0.2884727
#>  [6,] 0.1991105 0.1991105 0.1991105
#>  [7,] 0.7085211 0.7085211 0.7085211
#>  [8,] 1.7863407 1.7863407 1.7863407
#>  [9,] 0.5141226 0.5141226 0.5141226
#> [10,] 0.6723677 0.6723677 0.6723677
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>            [,1]       [,2]        [,3]
#>  [1,] 0.1878509 0.01371725 0.275706358
#>  [2,] 0.2149907 0.63359745 1.131105975
#>  [3,] 3.2106743 0.84354614 0.913902484
#>  [4,] 0.4146610 1.71555799 0.625722709
#>  [5,] 0.9636987 0.89796193 0.230890257
#>  [6,] 0.7205395 1.41413306 0.039698179
#>  [7,] 0.0344115 0.14056491 0.006265534
#>  [8,] 0.9107777 0.23480728 0.341910758
#>  [9,] 2.7780901 0.48035927 1.196085719
#> [10,] 1.7939688 1.40178846 0.278472083

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.34847948 0.1752492 0.5124481
#>  [2,] 0.12103403 0.1210340 0.1210340
#>  [3,] 1.42352108 0.2907384 1.3601563
#>  [4,] 1.87260101 5.3214488 1.4313064
#>  [5,] 0.16124018 0.9168566 0.2284265
#>  [6,] 1.39812642 0.3423706 0.1685531
#>  [7,] 0.53978912 0.7684020 4.3649152
#>  [8,] 2.67698840 2.7724064 0.1273391
#>  [9,] 1.99506054 2.1710502 2.4128817
#> [10,] 0.08676504 1.2744983 1.9057756
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05252337 0.05252337 0.05252337
#>  [2,] 1.48052008 1.48052008 1.48052008
#>  [3,] 2.78780778 2.78780778 2.78780778
#>  [4,] 1.16381211 1.16381211 1.16381211
#>  [5,] 0.15190351 0.15190351 0.15190351
#>  [6,] 1.39330756 1.39330756 1.39330756
#>  [7,] 0.06245884 0.06245884 0.06245884
#>  [8,] 2.23280870 2.23280870 2.23280870
#>  [9,] 0.97388135 0.97388135 0.97388135
#> [10,] 3.72598838 3.72598838 3.72598838
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 1.70882004 1.5531211 2.14114922
#>  [2,] 2.28503990 1.4639315 1.87531591
#>  [3,] 1.45049801 0.6702189 2.02061804
#>  [4,] 2.15071747 0.6673194 0.01536923
#>  [5,] 1.99884858 0.7895211 0.47641416
#>  [6,] 0.98433098 0.3069561 1.42981763
#>  [7,] 0.97889298 0.5560840 1.54107114
#>  [8,] 1.22499512 4.0597908 3.47143646
#>  [9,] 0.07465866 0.3971728 0.37907604
#> [10,] 0.49168864 2.1762466 0.23710540

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.31624710 1.1156528 2.3035238
#>  [2,] 2.39392919 0.2797974 1.8215817
#>  [3,] 0.97670072 0.8392064 0.5093417
#>  [4,] 2.06625294 2.6038396 0.3597017
#>  [5,] 2.66416209 0.3312795 2.9712182
#>  [6,] 0.05716018 0.4435403 0.9146855
#>  [7,] 1.44018977 7.5552256 1.3127263
#>  [8,] 1.15800377 1.0898795 1.1580038
#>  [9,] 0.35786936 0.1607178 5.6266366
#> [10,] 3.43366458 1.3423346 1.1505244
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29180010 0.29180010 0.29180010
#>  [2,] 0.55060526 0.55060526 0.55060526
#>  [3,] 0.61504099 0.61504099 0.61504099
#>  [4,] 0.77597643 0.77597643 0.77597643
#>  [5,] 0.34104636 0.34104636 0.34104636
#>  [6,] 0.47233931 0.47233931 0.47233931
#>  [7,] 0.07189076 0.07189076 0.07189076
#>  [8,] 0.78753701 0.78753701 0.78753701
#>  [9,] 1.02963067 1.02963067 1.02963067
#> [10,] 0.29645036 0.29645036 0.29645036
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5834521 1.12210573 0.9894430
#>  [2,] 0.2548871 0.83997851 0.1706210
#>  [3,] 2.4256510 0.50496453 1.4506189
#>  [4,] 0.5967606 0.78306277 4.5273291
#>  [5,] 1.1431781 1.51169498 0.5744406
#>  [6,] 0.6899477 0.58167870 0.6403392
#>  [7,] 0.6068420 1.04447471 1.8588868
#>  [8,] 0.1075834 1.37037001 0.1538278
#>  [9,] 1.6171254 0.85790483 1.4035309
#> [10,] 0.9520509 0.02318729 1.6525493

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0714612 1.0714612 1.0714612
#>  [2,] 0.1872608 0.1872608 0.1872608
#>  [3,] 3.2741505 2.8513893 1.9663071
#>  [4,] 0.7894793 2.9155438 2.3394374
#>  [5,] 0.9253552 0.9253552 0.9253552
#>  [6,] 0.4925571 0.1135671 0.8099130
#>  [7,] 4.2240382 4.9683918 2.2033778
#>  [8,] 0.4302209 0.5055874 0.3986492
#>  [9,] 0.6337646 0.6337646 0.6337646
#> [10,] 3.4192394 2.1265418 1.4486124
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06328117 0.06328117 0.06328117
#>  [2,] 0.02442871 0.02442871 0.02442871
#>  [3,] 1.33754554 1.33754554 1.33754554
#>  [4,] 0.49571947 0.49571947 0.49571947
#>  [5,] 1.82543526 1.82543526 1.82543526
#>  [6,] 0.81059119 0.81059119 0.81059119
#>  [7,] 0.27327994 0.27327994 0.27327994
#>  [8,] 1.14927077 1.14927077 1.14927077
#>  [9,] 0.54658110 0.54658110 0.54658110
#> [10,] 0.31404279 0.31404279 0.31404279
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.01406866 1.7290297 2.57098345
#>  [2,] 0.78885845 0.2122963 1.15066550
#>  [3,] 1.76545975 0.1435839 0.46621210
#>  [4,] 1.06143610 2.3745648 1.29864246
#>  [5,] 0.21635225 0.3241390 0.42093628
#>  [6,] 1.45847077 1.1331808 1.02941036
#>  [7,] 1.47298520 2.0166518 0.95974465
#>  [8,] 0.08714285 0.4146285 0.44429554
#>  [9,] 0.52029831 0.1805874 0.04484223
#> [10,] 0.19984306 0.3965980 2.72441901

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.821534429 0.5124372 0.2583252
#>  [2,] 1.019742675 1.2734965 0.3492365
#>  [3,] 3.404143413 1.6962472 0.6465533
#>  [4,] 1.612622545 0.5061458 0.9668485
#>  [5,] 0.135290418 3.2912002 1.2630118
#>  [6,] 0.312865938 2.1731272 0.1725535
#>  [7,] 0.002873623 0.3685258 0.8711777
#>  [8,] 1.168385807 3.5475997 2.1795702
#>  [9,] 0.104827722 0.1048277 0.1048277
#> [10,] 1.405632658 2.2074234 1.2362011
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3475977 0.3475977 0.3475977
#>  [2,] 0.8124254 0.8124254 0.8124254
#>  [3,] 2.9927674 2.9927674 2.9927674
#>  [4,] 1.2807103 1.2807103 1.2807103
#>  [5,] 3.2217757 3.2217757 3.2217757
#>  [6,] 0.9113201 0.9113201 0.9113201
#>  [7,] 2.8129886 2.8129886 2.8129886
#>  [8,] 2.8191942 2.8191942 2.8191942
#>  [9,] 0.2007977 0.2007977 0.2007977
#> [10,] 0.8598084 0.8598084 0.8598084
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6329036 0.28635171 0.3457225
#>  [2,] 1.7567939 0.63207923 1.6847778
#>  [3,] 0.7927992 0.13517836 0.1654133
#>  [4,] 0.7552047 0.61927440 0.2837631
#>  [5,] 0.2458031 0.44461441 0.8037200
#>  [6,] 0.2257716 0.07462834 1.1780779
#>  [7,] 0.9288770 1.36782625 2.0292580
#>  [8,] 0.9580277 2.36979298 0.2793134
#>  [9,] 1.9783996 0.31663106 1.0738306
#> [10,] 0.1296050 3.02591488 0.5392239

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.495246421 0.49524642 0.49524642
#>  [2,] 1.091646385 0.06264087 2.61933100
#>  [3,] 1.226927136 1.56425653 3.99973257
#>  [4,] 0.102259469 0.10225947 0.10225947
#>  [5,] 1.343486674 3.07448906 1.90042774
#>  [6,] 0.036980025 3.07463455 0.45704542
#>  [7,] 2.181726706 2.90458416 0.09273702
#>  [8,] 1.098438036 0.28575555 1.14025700
#>  [9,] 1.551537991 0.03320203 1.55153799
#> [10,] 0.006958139 2.04205784 1.33092671
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1028152 1.1028152 1.1028152
#>  [2,] 0.2865518 0.2865518 0.2865518
#>  [3,] 3.1949871 3.1949871 3.1949871
#>  [4,] 2.4397070 2.4397070 2.4397070
#>  [5,] 3.5291137 3.5291137 3.5291137
#>  [6,] 2.3648887 2.3648887 2.3648887
#>  [7,] 0.9877169 0.9877169 0.9877169
#>  [8,] 1.8046324 1.8046324 1.8046324
#>  [9,] 0.3717459 0.3717459 0.3717459
#> [10,] 0.5029127 0.5029127 0.5029127
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>               [,1]        [,2]       [,3]
#>  [1,] 0.4538626351 1.757961823 2.62079398
#>  [2,] 0.5874731843 1.129197938 0.47010594
#>  [3,] 0.2255108892 0.419328163 0.25678809
#>  [4,] 0.0502985000 0.316406655 0.23362977
#>  [5,] 2.5230293319 0.007633708 0.08067816
#>  [6,] 0.4163208452 2.642199269 0.65609823
#>  [7,] 0.0657248856 0.618311155 0.53728624
#>  [8,] 1.6141530559 0.518356233 0.51198499
#>  [9,] 1.3546463008 2.744488489 1.11809805
#> [10,] 0.0001997865 0.288931870 1.29998752

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5484324 0.2804406 1.7567493
#>  [2,] 0.8590592 0.8590592 0.8590592
#>  [3,] 0.3234311 1.0218240 1.6335513
#>  [4,] 1.7806278 0.7603725 0.3653997
#>  [5,] 1.1302114 1.1302114 1.1302114
#>  [6,] 1.3112783 0.3695630 0.8063093
#>  [7,] 2.5656347 0.3079411 0.6703233
#>  [8,] 0.0383946 0.0383946 0.0383946
#>  [9,] 1.1268028 3.3020080 3.9483523
#> [10,] 2.8610265 1.8455450 2.8610265
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5311076 1.5311076 1.5311076
#>  [2,] 2.6902554 2.6902554 2.6902554
#>  [3,] 1.1125384 1.1125384 1.1125384
#>  [4,] 0.3918773 0.3918773 0.3918773
#>  [5,] 0.3288415 0.3288415 0.3288415
#>  [6,] 1.0673250 1.0673250 1.0673250
#>  [7,] 0.7302255 0.7302255 0.7302255
#>  [8,] 0.4554462 0.4554462 0.4554462
#>  [9,] 0.8499802 0.8499802 0.8499802
#> [10,] 0.5968806 0.5968806 0.5968806
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 1.5313827 0.5787146 0.95303691
#>  [2,] 0.4818984 0.6527007 4.23348993
#>  [3,] 0.5228494 3.7551730 1.28528036
#>  [4,] 0.9541545 3.3513600 0.40590671
#>  [5,] 0.8144350 1.1017749 0.63435869
#>  [6,] 0.2253953 0.6418872 0.24351728
#>  [7,] 3.1542269 1.7910727 0.16150912
#>  [8,] 0.8170281 3.3482412 1.34931697
#>  [9,] 2.3104671 0.6410992 1.21577880
#> [10,] 1.2123702 0.8642527 0.07220098

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>            [,1]         [,2]      [,3]
#>  [1,] 0.5000443 0.7690965361 0.9887410
#>  [2,] 0.7101837 0.1174770092 0.1174770
#>  [3,] 1.3454227 0.4701996762 1.3454227
#>  [4,] 0.7720751 0.7030787972 0.7694201
#>  [5,] 0.2540857 0.7427056372 1.0764855
#>  [6,] 2.1975160 0.4327537085 0.8567412
#>  [7,] 1.3799955 1.3799954692 0.1733102
#>  [8,] 1.1127653 1.1127653497 0.2479685
#>  [9,] 0.5732760 0.5732760228 0.5732760
#> [10,] 0.4695232 0.0008972489 0.9420100
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]         [,3]
#>  [1,] 0.88767515 0.88767515 1.0554497597
#>  [2,] 0.96851773 0.96851773 0.8278214572
#>  [3,] 1.24110658 1.36593683 0.2323884575
#>  [4,] 0.21416648 0.21416648 0.0005288636
#>  [5,] 1.16409701 0.83507360 0.8350735951
#>  [6,] 0.38976742 0.38976742 0.3897674185
#>  [7,] 0.01591154 0.70369884 2.1099953486
#>  [8,] 0.31522183 0.70860676 0.7086067616
#>  [9,] 0.38101177 0.09992473 0.8860880744
#> [10,] 0.36249630 0.59856083 0.3624962966
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.02442617 0.024426173 0.02442617
#>  [2,] 0.32612358 0.326123575 0.32612358
#>  [3,] 2.31821252 2.318212520 0.67414361
#>  [4,] 2.63579658 1.607278756 1.68784668
#>  [5,] 1.43555554 0.382645816 2.67040244
#>  [6,] 1.68489577 0.802780362 1.68489577
#>  [7,] 1.43376719 1.016857726 1.21231081
#>  [8,] 0.08135902 0.800352663 0.48189563
#>  [9,] 2.14510472 0.246179872 1.36300758
#> [10,] 0.07770146 0.006804007 0.10855400
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4183366 0.5860298 0.09892717
#>  [2,] 0.6344535 0.3287003 0.54320326
#>  [3,] 0.4698760 0.4698760 0.46987604
#>  [4,] 0.5413573 1.1492519 0.16162350
#>  [5,] 0.2874233 0.2874233 0.28742334
#>  [6,] 0.5583636 1.2714316 0.60374090
#>  [7,] 2.1490596 1.9959802 0.68048217
#>  [8,] 0.6057257 0.3788257 0.60572568
#>  [9,] 0.1175898 0.5132422 0.01254579
#> [10,] 0.2599048 0.2599048 0.25990475
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8193081 0.4021496 1.2894830
#>  [2,] 0.7532066 0.1010567 0.1010567
#>  [3,] 0.2337851 0.2530022 1.5094504
#>  [4,] 0.1569829 0.3924594 0.1569829
#>  [5,] 0.3674490 0.3385673 0.7098336
#>  [6,] 1.0778127 0.5296200 1.0778127
#>  [7,] 2.3825507 0.3749242 0.3729938
#>  [8,] 2.5383284 0.2023501 1.2040411
#>  [9,] 0.3436124 0.4355059 0.4684442
#> [10,] 1.0149743 0.5715957 0.2203953

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>              [,1]       [,2]      [,3]
#>  [1,] 4.644073514 0.19879549 1.6095301
#>  [2,] 0.006641266 0.16986052 0.3126564
#>  [3,] 1.134804917 0.25180957 1.1348049
#>  [4,] 2.139161606 2.13916161 0.2579638
#>  [5,] 0.653541384 0.03368171 1.1349944
#>  [6,] 0.712279599 0.67758974 0.9702494
#>  [7,] 0.381904006 0.38190401 0.3819040
#>  [8,] 2.041350998 0.11879883 0.4251084
#>  [9,] 0.252479336 1.04755929 1.2036813
#> [10,] 0.039097441 1.05849145 1.7970742
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5523974 0.07217267 0.8838683
#>  [2,] 0.6225345 0.62253446 0.6225345
#>  [3,] 0.7455362 2.77108662 0.0432745
#>  [4,] 1.4909838 1.49098383 0.6824325
#>  [5,] 0.8045118 3.01771348 2.0571771
#>  [6,] 1.3883826 0.03707343 0.8336338
#>  [7,] 0.8159264 0.64553553 0.8159264
#>  [8,] 0.7270036 0.72700356 0.2922239
#>  [9,] 0.6942922 1.19312217 1.1138897
#> [10,] 0.6120234 0.70128740 0.9199246
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 3.25616992 1.08058601 0.79476442
#>  [2,] 1.36169474 2.13297456 1.91501701
#>  [3,] 0.81634302 0.81634302 0.81634302
#>  [4,] 2.54250745 1.71510100 1.46924612
#>  [5,] 1.32421540 1.83769027 2.73696181
#>  [6,] 2.97724957 0.06192255 0.45465066
#>  [7,] 0.87099552 5.21758085 0.08233002
#>  [8,] 5.25688446 1.16524450 1.90785595
#>  [9,] 2.72891721 0.30398814 0.58391070
#> [10,] 0.08018541 1.73629411 1.83985999
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 4.5249354 1.3804748 0.4910309
#>  [2,] 0.2046196 0.7315946 1.8276376
#>  [3,] 3.9566052 0.5952909 0.7769195
#>  [4,] 1.1680277 1.8123755 0.9393930
#>  [5,] 0.4350019 0.5678479 1.7761740
#>  [6,] 0.4462020 0.4730607 0.0180714
#>  [7,] 2.6754779 1.9941139 2.6754779
#>  [8,] 0.1382371 0.1382371 0.1382371
#>  [9,] 0.5102972 4.7629791 2.6000489
#> [10,] 1.3117044 1.1494233 0.4228395
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.268816451 1.2688165 0.5381924
#>  [2,] 1.017536886 2.3182009 1.0175369
#>  [3,] 0.118518692 0.7136636 0.5065690
#>  [4,] 1.089750833 1.0897508 1.0897508
#>  [5,] 0.008549559 0.3814760 0.3814760
#>  [6,] 0.096551157 2.6395441 0.8726052
#>  [7,] 1.805941116 6.8846758 1.1655638
#>  [8,] 0.928528037 0.6640017 0.7699240
#>  [9,] 3.677076493 0.4985390 0.2631209
#> [10,] 3.374837359 2.5507102 0.9468007

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]      [,3]
#>  [1,] 2.23362902 3.55748448 1.6722516
#>  [2,] 0.06584688 1.60588330 2.4799177
#>  [3,] 0.71498154 0.05245324 0.7606094
#>  [4,] 0.93007246 0.66565091 0.2139954
#>  [5,] 0.35674512 0.71959285 0.3567451
#>  [6,] 0.27616105 0.99948791 0.1465008
#>  [7,] 0.19587050 0.71096107 0.5837770
#>  [8,] 0.95693083 0.57267150 0.4247751
#>  [9,] 0.70399479 1.14877086 0.9658484
#> [10,] 0.28110542 0.28110542 0.2811054
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40265501 0.40265501 0.40265501
#>  [2,] 0.03465087 0.08208112 0.08208112
#>  [3,] 0.31848462 0.27094171 0.31848462
#>  [4,] 0.03926416 0.03926416 0.03926416
#>  [5,] 0.02578607 0.02578607 0.02578607
#>  [6,] 0.59690557 0.57101835 0.59690557
#>  [7,] 0.04682787 0.04682787 0.04682787
#>  [8,] 1.17864289 1.08582967 0.87823097
#>  [9,] 1.37793690 0.10231353 1.08433261
#> [10,] 0.45732731 0.45732731 0.45732731
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.43607085 0.43607085 0.43607085
#>  [2,] 0.05608407 0.66929278 0.88452007
#>  [3,] 0.09568230 0.09568230 0.09568230
#>  [4,] 0.62004009 0.62004009 0.62004009
#>  [5,] 1.62834885 1.62834885 1.62834885
#>  [6,] 0.05980169 0.05322321 0.05980169
#>  [7,] 0.32480469 0.32480469 0.32480469
#>  [8,] 1.31922236 1.31922236 0.25849870
#>  [9,] 0.12891339 0.11967887 0.11967887
#> [10,] 1.54802964 0.17986917 0.44777948
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33525128 0.85069545 2.24528615
#>  [2,] 0.71704911 0.71704911 0.15531694
#>  [3,] 0.83392257 0.19088589 0.19088589
#>  [4,] 0.44517747 0.44517747 0.44517747
#>  [5,] 0.78026847 0.20690994 0.78026847
#>  [6,] 0.01044286 0.01044286 0.01044286
#>  [7,] 0.06062064 0.06062064 0.06062064
#>  [8,] 0.01050312 0.01050312 0.01050312
#>  [9,] 0.41050883 0.41050883 0.48384940
#> [10,] 0.39476155 0.17508560 0.17508560
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08982778 0.34129381 0.32447330
#>  [2,] 0.33186735 0.09363992 0.67899656
#>  [3,] 0.17973971 0.17973971 0.21526540
#>  [4,] 1.00521048 0.34041640 0.26833369
#>  [5,] 0.04669707 0.04669707 0.04669707
#>  [6,] 0.64190071 0.64190071 0.12485694
#>  [7,] 2.47876642 0.55702328 0.43436366
#>  [8,] 0.48019741 0.59576269 0.48019741
#>  [9,] 0.14560132 0.14560132 0.14560132
#> [10,] 0.25597501 0.01823723 0.25597501

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26492283 0.27017069 0.27017069
#>  [2,] 0.19303551 0.19303551 0.06732865
#>  [3,] 0.07417416 0.27953465 0.40044141
#>  [4,] 0.14791748 0.14791748 0.14791748
#>  [5,] 0.49099624 0.76505469 0.14317699
#>  [6,] 1.51287920 1.62193161 0.95398707
#>  [7,] 0.87188346 1.69608355 1.40081601
#>  [8,] 0.63886978 0.52772796 1.32365069
#>  [9,] 0.33889314 0.03507425 0.33889314
#> [10,] 1.63057759 0.16620259 0.16620259
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23747845 0.03780859 0.31970250
#>  [2,] 0.17757901 0.24041433 0.56042946
#>  [3,] 0.80858705 0.13268880 0.13268880
#>  [4,] 0.65539111 0.64380870 0.01900476
#>  [5,] 0.11075194 0.64777481 0.64777481
#>  [6,] 0.03057248 0.31146524 0.31146524
#>  [7,] 1.09783591 1.09783591 0.48903355
#>  [8,] 0.47983686 0.11972054 0.27567941
#>  [9,] 0.24189282 0.24189282 0.13090310
#> [10,] 0.09608951 0.92589950 0.61799284
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.89969196 0.20461468 0.28890011
#>  [2,] 1.90105094 0.07782982 0.09285023
#>  [3,] 0.16643334 0.54591620 0.16643334
#>  [4,] 0.03300034 0.03300034 0.03300034
#>  [5,] 0.03177995 0.26044296 0.03177995
#>  [6,] 0.14399394 0.14399394 0.14399394
#>  [7,] 0.60056254 0.45930518 0.60056254
#>  [8,] 0.49532770 0.32310301 0.52252195
#>  [9,] 0.33934539 0.25152073 0.58867900
#> [10,] 0.39550392 0.39550392 0.58957882
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03559330 0.84566730 2.13209110
#>  [2,] 0.03760616 0.03760616 0.03760616
#>  [3,] 0.22879111 0.01777325 0.11972274
#>  [4,] 0.38449760 0.70314351 0.21940718
#>  [5,] 0.36373282 0.27523919 0.36373282
#>  [6,] 1.17750660 0.33478896 1.17750660
#>  [7,] 0.24636688 0.24636688 0.12304575
#>  [8,] 0.26451457 0.26451457 0.77446057
#>  [9,] 0.50739865 0.86736451 0.86736451
#> [10,] 0.01291547 0.01291547 0.01291547

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09990706 0.09990706 0.09990706
#>  [2,] 1.18751710 1.18751710 0.32652324
#>  [3,] 0.15312883 0.11468434 0.42503278
#>  [4,] 0.71801174 0.18652329 0.14586593
#>  [5,] 0.44913760 0.33357996 0.39796410
#>  [6,] 0.53665520 1.10079940 0.09528513
#>  [7,] 0.05893225 0.09365363 0.05893225
#>  [8,] 0.12108657 0.12108657 0.12108657
#>  [9,] 0.37523796 0.12842065 0.02068079
#> [10,] 0.03489199 0.03489199 0.05744578
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06671751 0.13832441 0.02653098
#>  [2,] 0.36780561 0.20557277 1.09331829
#>  [3,] 1.06839318 1.09419744 0.28637800
#>  [4,] 0.39102399 0.06738241 0.56644234
#>  [5,] 0.34415599 0.34415599 0.42441077
#>  [6,] 0.01221163 0.01221163 0.15616976
#>  [7,] 0.59119280 1.36118079 0.55644257
#>  [8,] 0.02361846 0.40107328 0.22972495
#>  [9,] 0.01664588 0.50416672 0.52371595
#> [10,] 0.05713009 1.19681080 1.43417501
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.09357567 0.14354286 0.113338100
#>  [2,] 0.03928740 2.27228048 1.743872076
#>  [3,] 0.17688623 0.23866408 0.529097090
#>  [4,] 0.03491004 0.08867294 0.034910037
#>  [5,] 0.11807764 0.17564678 0.147118880
#>  [6,] 0.13145030 0.39020430 0.078375350
#>  [7,] 0.10777553 1.39806583 0.107775533
#>  [8,] 0.39078829 0.34100153 0.005165642
#>  [9,] 0.50330970 0.68422935 0.500176196
#> [10,] 0.41563831 0.50056413 0.500564131
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.277207565 0.365208548 0.497789659
#>  [2,] 0.840237958 0.369083973 0.962662584
#>  [3,] 0.347331892 0.672377588 0.339239192
#>  [4,] 0.378654682 0.824878566 0.824878566
#>  [5,] 0.091689632 0.091689632 0.091689632
#>  [6,] 0.149681060 0.663300370 0.695700775
#>  [7,] 0.085402337 0.085402337 0.041565843
#>  [8,] 0.093187721 0.093187721 0.414335720
#>  [9,] 0.439047172 0.081474056 0.081474056
#> [10,] 0.004430777 0.004430777 0.004430777

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.389908494 0.582403670 1.339686339
#>  [2,] 0.316287568 0.167020384 0.719310377
#>  [3,] 0.009159823 0.009159823 0.009159823
#>  [4,] 0.375060964 0.899617981 0.354257363
#>  [5,] 0.039249371 0.054516911 0.142273396
#>  [6,] 1.223447597 0.788383667 0.236000390
#>  [7,] 0.066216414 0.309138972 0.309138972
#>  [8,] 0.682377492 0.711347460 0.537624441
#>  [9,] 0.986397967 0.092733293 0.986397967
#> [10,] 0.227786110 0.437341050 0.041089882
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.253975080 0.25397508 0.08641676
#>  [2,] 0.063711740 0.06371174 0.06371174
#>  [3,] 0.368019000 0.15667767 0.15667767
#>  [4,] 0.454427935 0.29766581 0.25068292
#>  [5,] 0.049562945 0.04956294 0.04956294
#>  [6,] 0.082091578 0.52895286 0.77961031
#>  [7,] 0.187827462 0.32391747 0.14206358
#>  [8,] 0.004712740 0.38009169 0.10559835
#>  [9,] 0.295765279 1.04019360 1.04019360
#> [10,] 0.003217908 0.06869231 0.06869231
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.56111396 0.5611140 0.12236885
#>  [2,] 0.33396326 0.5071908 0.29709955
#>  [3,] 0.63951442 1.4532034 0.80850638
#>  [4,] 0.54045877 0.4429401 0.54045877
#>  [5,] 0.10467250 0.4882745 0.10467250
#>  [6,] 0.03572373 0.4258693 0.03572373
#>  [7,] 0.48310404 0.3699971 0.36999709
#>  [8,] 0.03129302 0.1795510 1.08993616
#>  [9,] 0.17829437 0.2210092 0.17829437
#> [10,] 0.30743044 1.4310605 1.43106051
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>              [,1]       [,2]         [,3]
#>  [1,] 0.026788966 0.02678897 0.0267889657
#>  [2,] 0.433944196 0.44487282 0.4339441965
#>  [3,] 0.065681035 0.30299954 0.0656810352
#>  [4,] 0.003637087 0.01277618 0.0036370868
#>  [5,] 0.250514448 0.19672533 0.0072952061
#>  [6,] 0.484771558 0.48477156 0.4847715582
#>  [7,] 0.823129427 0.57864453 0.5294912785
#>  [8,] 0.134617956 0.06541364 0.1161885624
#>  [9,] 0.174173192 0.00144776 0.0245517185
#> [10,] 0.495585614 0.51264786 0.0002988644