Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1026381 3.9138918 0.14415921
#>  [2,] 5.1456770 0.2260928 0.78938536
#>  [3,] 4.6841205 0.5989541 1.30449656
#>  [4,] 0.2524410 8.0840517 0.56590807
#>  [5,] 0.4153921 2.3922883 0.92282570
#>  [6,] 0.4919044 0.4919044 0.01625241
#>  [7,] 0.4610188 1.2541720 0.60039349
#>  [8,] 0.3517068 8.1854134 0.68168569
#>  [9,] 0.4249358 0.9356699 0.19333029
#> [10,] 1.0784114 0.3893275 0.11642828
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.99615079 1.99615079 1.99615079
#>  [2,] 0.83957617 0.83957617 0.83957617
#>  [3,] 4.01677853 4.01677853 4.01677853
#>  [4,] 0.60538744 0.60538744 0.60538744
#>  [5,] 3.64065630 3.64065630 3.64065630
#>  [6,] 0.03554049 0.03554049 0.03554049
#>  [7,] 1.28549782 1.28549782 1.28549782
#>  [8,] 0.10918988 0.10918988 0.10918988
#>  [9,] 0.02512938 0.02512938 0.02512938
#> [10,] 0.25097327 0.25097327 0.25097327
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]      [,2]        [,3]
#>  [1,] 0.66888000 3.2400292 1.079227411
#>  [2,] 0.02232965 0.6649997 0.041756979
#>  [3,] 1.68542834 2.2553783 0.003664206
#>  [4,] 0.78963914 2.2379560 4.108893459
#>  [5,] 0.05086199 0.7450793 0.148062612
#>  [6,] 0.33305781 0.3854124 0.643591581
#>  [7,] 1.38201679 1.4208887 0.541001444
#>  [8,] 1.41248423 0.2159214 1.567995999
#>  [9,] 0.31010714 0.3779324 0.102438880
#> [10,] 0.18270528 1.2635582 1.382442016

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.67652200 1.67652200 1.67652200
#>  [2,] 0.26439058 3.17923360 1.43451620
#>  [3,] 1.82340115 1.82340115 0.73042917
#>  [4,] 0.29692786 0.42130613 1.25215711
#>  [5,] 1.09307657 1.09307657 0.28389824
#>  [6,] 0.89825944 0.62271710 1.04398765
#>  [7,] 1.56549590 1.91230840 0.84554201
#>  [8,] 0.07280148 0.07280148 0.07280148
#>  [9,] 6.65373671 4.08811335 1.60424314
#> [10,] 4.36291235 0.29961887 0.22711910
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1650778 0.1650778 0.1650778
#>  [2,] 0.4586102 0.4586102 0.4586102
#>  [3,] 0.1785768 0.1785768 0.1785768
#>  [4,] 0.6075957 0.6075957 0.6075957
#>  [5,] 0.4822989 0.4822989 0.4822989
#>  [6,] 0.4579231 0.4579231 0.4579231
#>  [7,] 2.1629708 2.1629708 2.1629708
#>  [8,] 1.6716164 1.6716164 1.6716164
#>  [9,] 1.7548187 1.7548187 1.7548187
#> [10,] 1.5861347 1.5861347 1.5861347
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>            [,1]        [,2]       [,3]
#>  [1,] 2.5400459 0.287327478 0.74859673
#>  [2,] 0.2061817 0.651840119 1.31233696
#>  [3,] 0.2713474 0.200729685 0.08125999
#>  [4,] 0.2309142 5.543675707 1.57083220
#>  [5,] 3.4278932 0.923433110 0.38208651
#>  [6,] 0.3346774 1.146189919 2.18450717
#>  [7,] 0.4118138 3.325845115 1.34454275
#>  [8,] 0.5451447 0.007727476 0.68396861
#>  [9,] 1.6588058 0.064873227 1.65459692
#> [10,] 0.5022606 0.080930539 0.22901656

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.8032951 0.31173754 1.8032951
#>  [2,] 0.0730274 0.07302740 0.0730274
#>  [3,] 0.4107596 0.02386493 0.1890562
#>  [4,] 0.7698875 1.38812972 0.9804409
#>  [5,] 1.0157217 0.71614070 0.7962215
#>  [6,] 1.8976349 0.05651613 1.8976349
#>  [7,] 1.7149836 3.96013325 2.9010324
#>  [8,] 1.7766273 1.28803143 1.7766273
#>  [9,] 0.5484935 4.90205664 0.2336172
#> [10,] 1.7072514 1.11120862 1.6102101
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 3.455699861 3.455699861 3.455699861
#>  [2,] 1.027290683 1.027290683 1.027290683
#>  [3,] 0.480475830 0.480475830 0.480475830
#>  [4,] 0.710442431 0.710442431 0.710442431
#>  [5,] 1.309387114 1.309387114 1.309387114
#>  [6,] 0.568870038 0.568870038 0.568870038
#>  [7,] 0.675621448 0.675621448 0.675621448
#>  [8,] 0.284964791 0.284964791 0.284964791
#>  [9,] 0.003631826 0.003631826 0.003631826
#> [10,] 0.065495228 0.065495228 0.065495228
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.79852010 0.09609042 1.10837497
#>  [2,] 0.68096611 1.49483490 0.62954679
#>  [3,] 1.10575382 0.79396489 0.64813004
#>  [4,] 0.04604583 2.80873929 0.18616187
#>  [5,] 0.43141504 0.02191498 4.05246138
#>  [6,] 0.64574368 0.21841614 0.01915976
#>  [7,] 0.52918876 0.32430003 1.15417011
#>  [8,] 2.70977133 0.59467535 1.72816877
#>  [9,] 1.42865051 1.68103236 0.02326190
#> [10,] 0.67827413 0.15784929 0.95253748

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6313008 0.49621440 1.2299004
#>  [2,] 0.2600332 0.48820553 2.3749202
#>  [3,] 1.6128546 0.08295252 0.8088924
#>  [4,] 0.1030613 2.80914974 4.6176175
#>  [5,] 0.6748604 2.45417804 0.0830072
#>  [6,] 1.2262019 1.22620186 1.2262019
#>  [7,] 0.2231910 4.74199344 7.0293778
#>  [8,] 1.0131376 1.03109057 0.9924670
#>  [9,] 2.2979887 1.91428548 1.2941012
#> [10,] 0.8289209 0.33575710 0.7119122
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2557984 1.2557984 1.2557984
#>  [2,] 1.0248125 1.0248125 1.0248125
#>  [3,] 1.1773681 1.1773681 1.1773681
#>  [4,] 1.4564832 1.4564832 1.4564832
#>  [5,] 0.9900739 0.9900739 0.9900739
#>  [6,] 0.3912862 0.3912862 0.3912862
#>  [7,] 0.2386054 0.2386054 0.2386054
#>  [8,] 1.6816183 1.6816183 1.6816183
#>  [9,] 1.4766458 1.4766458 1.4766458
#> [10,] 1.9255554 1.9255554 1.9255554
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4216063 2.4972959 0.8475900
#>  [2,] 1.0211348 0.3825729 4.1623413
#>  [3,] 2.3221889 1.5593268 0.1592743
#>  [4,] 0.9385671 3.5493556 0.9880903
#>  [5,] 1.1776201 1.0942023 0.5988797
#>  [6,] 0.4051132 1.0934674 0.9101672
#>  [7,] 0.5601579 1.2562353 0.7581986
#>  [8,] 2.1458580 2.3721264 3.0885302
#>  [9,] 0.3778856 0.8123082 2.4494867
#> [10,] 0.8062492 3.1143400 1.8359071

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.67735556 0.1727306 0.3316494
#>  [2,] 1.83955893 1.4333432 0.3610398
#>  [3,] 2.97393266 2.9733578 0.3244373
#>  [4,] 3.63073488 1.1159330 2.5832872
#>  [5,] 0.40296934 0.4029693 0.4029693
#>  [6,] 0.29921974 0.2992197 0.2992197
#>  [7,] 0.04777417 1.8969934 4.0981271
#>  [8,] 0.44373459 0.8754882 1.8287978
#>  [9,] 1.41190295 2.0622478 1.7036062
#> [10,] 0.93592805 5.4615486 0.4870836
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4021478 1.4021478 1.4021478
#>  [2,] 0.1980864 0.1980864 0.1980864
#>  [3,] 0.7139394 0.7139394 0.7139394
#>  [4,] 0.8636322 0.8636322 0.8636322
#>  [5,] 0.4192895 0.4192895 0.4192895
#>  [6,] 0.2099730 0.2099730 0.2099730
#>  [7,] 1.9167185 1.9167185 1.9167185
#>  [8,] 0.4552394 0.4552394 0.4552394
#>  [9,] 0.8430326 0.8430326 0.8430326
#> [10,] 0.8944109 0.8944109 0.8944109
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]        [,2]      [,3]
#>  [1,] 0.02899117 0.986450026 0.6309242
#>  [2,] 0.64352891 0.974731795 0.9050621
#>  [3,] 0.04098699 0.183328693 2.6974664
#>  [4,] 2.90224426 0.019045015 2.5423571
#>  [5,] 2.23820027 0.156685091 1.1900117
#>  [6,] 0.26338891 0.001722662 0.5252757
#>  [7,] 0.39938635 0.164322015 1.3109189
#>  [8,] 1.36751045 3.511481890 0.2870797
#>  [9,] 1.02689352 0.415540113 0.2461495
#> [10,] 0.46087076 0.226511095 1.0043969

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.818711451 0.8187115 0.7761885
#>  [2,] 0.158009740 0.1947344 1.0924320
#>  [3,] 1.694830360 1.1523496 0.7930637
#>  [4,] 1.034389870 0.8542497 3.9627135
#>  [5,] 0.137205073 2.5655891 2.5655891
#>  [6,] 3.387288409 3.2096765 1.9070254
#>  [7,] 0.540447536 0.8008352 2.6700099
#>  [8,] 0.009653494 0.1892579 0.1892579
#>  [9,] 0.683544388 0.6835444 0.6835444
#> [10,] 1.209977880 0.6418915 1.3603336
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3447758 0.3447758 0.3447758
#>  [2,] 0.6179976 0.6179976 0.6179976
#>  [3,] 0.2228542 0.2228542 0.2228542
#>  [4,] 1.3163268 1.3163268 1.3163268
#>  [5,] 0.4199397 0.4199397 0.4199397
#>  [6,] 2.8981631 2.8981631 2.8981631
#>  [7,] 4.8593489 4.8593489 4.8593489
#>  [8,] 1.0938821 1.0938821 1.0938821
#>  [9,] 1.7321413 1.7321413 1.7321413
#> [10,] 5.2684124 5.2684124 5.2684124
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 1.24359433 2.08985304 1.3845366
#>  [2,] 1.09650192 0.83940728 2.1981447
#>  [3,] 0.25565434 0.00109965 0.2674783
#>  [4,] 0.48303591 0.44487768 0.3282372
#>  [5,] 0.07220362 1.33081315 1.1160243
#>  [6,] 0.13266648 0.08932323 0.1605280
#>  [7,] 0.26806504 0.95483341 0.2365866
#>  [8,] 0.71047389 0.14947808 0.1691470
#>  [9,] 2.25881127 1.00217426 6.8365028
#> [10,] 1.42656733 2.03860064 0.4494433

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>             [,1]        [,2]         [,3]
#>  [1,] 1.37893838 2.615684066  0.971759545
#>  [2,] 1.60896877 1.139307602  1.420547526
#>  [3,] 1.22629259 0.619156006  1.226292595
#>  [4,] 0.08090256 0.008642526  0.080902561
#>  [5,] 4.87489563 2.010165419  7.369209747
#>  [6,] 4.36959089 1.600723977  0.002212064
#>  [7,] 0.05487006 0.768626074  0.068995710
#>  [8,] 1.59923583 6.629214108 12.264703148
#>  [9,] 0.80349506 1.877034582  2.052067986
#> [10,] 0.33053213 1.165387348  0.593015893
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2896111 0.2896111 0.2896111
#>  [2,] 3.4297085 3.4297085 3.4297085
#>  [3,] 2.2035371 2.2035371 2.2035371
#>  [4,] 0.6723335 0.6723335 0.6723335
#>  [5,] 1.0801322 1.0801322 1.0801322
#>  [6,] 0.0236073 0.0236073 0.0236073
#>  [7,] 9.8552141 9.8552141 9.8552141
#>  [8,] 1.7017618 1.7017618 1.7017618
#>  [9,] 1.8854890 1.8854890 1.8854890
#> [10,] 0.5593475 0.5593475 0.5593475
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2678140 0.92649128 0.7942629
#>  [2,] 0.5470553 0.33402325 0.5835451
#>  [3,] 1.0881875 0.76114390 0.1874700
#>  [4,] 0.5069130 0.96713034 0.2247147
#>  [5,] 0.7297576 1.16821016 0.3345287
#>  [6,] 1.8212722 0.08466452 1.3804342
#>  [7,] 0.4702676 0.22792710 0.4851389
#>  [8,] 1.0186848 1.21840995 0.3622838
#>  [9,] 0.6353716 0.05813500 2.4336618
#> [10,] 2.5596885 2.10798880 1.0028356

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1015247 2.2287885 1.0734704
#>  [2,] 0.3197448 1.8970145 0.2992605
#>  [3,] 1.3219921 1.2721834 1.1126171
#>  [4,] 2.9741728 1.1991728 1.2809890
#>  [5,] 0.4502835 0.4502835 0.4502835
#>  [6,] 2.4064656 0.1778481 0.4186639
#>  [7,] 2.3390749 1.4505359 4.1730921
#>  [8,] 2.0377579 2.0377579 0.4290300
#>  [9,] 4.8627918 0.1960172 0.3496326
#> [10,] 1.0552787 0.6518838 1.1312962
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.97730276 0.97730276 0.97730276
#>  [2,] 1.69856348 1.69856348 1.69856348
#>  [3,] 1.12833803 1.12833803 1.12833803
#>  [4,] 1.81051479 1.81051479 1.81051479
#>  [5,] 1.03009865 1.03009865 1.03009865
#>  [6,] 0.08575415 0.08575415 0.08575415
#>  [7,] 0.71263500 0.71263500 0.71263500
#>  [8,] 0.80397401 0.80397401 0.80397401
#>  [9,] 0.61382308 0.61382308 0.61382308
#> [10,] 0.30718231 0.30718231 0.30718231
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4466138 0.3308010 0.5114535
#>  [2,] 1.8710983 0.1559379 0.7519214
#>  [3,] 0.1368556 0.5438709 3.1359962
#>  [4,] 1.2918148 1.4572910 2.0269844
#>  [5,] 0.6163992 0.2824468 1.8314468
#>  [6,] 2.2272320 1.0780475 1.2794410
#>  [7,] 0.2538597 0.2106055 0.8629944
#>  [8,] 0.3617938 0.5386387 0.1551681
#>  [9,] 0.4017386 0.5144342 4.8054090
#> [10,] 1.2030150 0.4005979 0.7400976

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4800844 0.4800844 0.48008439
#>  [2,] 0.8688472 0.8508156 1.56666767
#>  [3,] 1.9086180 0.2908205 0.17464278
#>  [4,] 0.2672477 0.2411029 0.01410154
#>  [5,] 0.5080690 0.5639801 0.31180353
#>  [6,] 0.2262329 0.2262329 0.22623286
#>  [7,] 4.5587389 0.9139470 0.57467990
#>  [8,] 1.5635943 3.6542059 5.68035474
#>  [9,] 3.5412716 1.1648164 0.92094926
#> [10,] 1.9995160 4.2353563 1.98481033
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5420501 0.5420501 0.5420501
#>  [2,] 3.4959064 3.4959064 3.4959064
#>  [3,] 0.7440041 0.7440041 0.7440041
#>  [4,] 0.9192801 0.9192801 0.9192801
#>  [5,] 4.6418849 4.6418849 4.6418849
#>  [6,] 0.2623523 0.2623523 0.2623523
#>  [7,] 1.8032540 1.8032540 1.8032540
#>  [8,] 1.0023583 1.0023583 1.0023583
#>  [9,] 0.3606152 0.3606152 0.3606152
#> [10,] 1.5577677 1.5577677 1.5577677
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>             [,1]       [,2]        [,3]
#>  [1,] 0.51854094 0.83271285 0.826851388
#>  [2,] 2.53917382 2.06705172 0.003538007
#>  [3,] 0.56104385 1.79756858 0.825069069
#>  [4,] 0.06756853 0.67998319 0.695461931
#>  [5,] 1.27928134 0.39937654 1.727850607
#>  [6,] 1.34135860 0.04800314 1.183665847
#>  [7,] 2.52101426 0.48290805 3.514548435
#>  [8,] 0.23384325 0.96243777 0.273958264
#>  [9,] 0.54280829 0.57000023 0.851707136
#> [10,] 1.24015626 2.37433903 1.285318558

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5927140 0.9961727 1.5942694
#>  [2,] 2.3421789 0.5409483 0.4420097
#>  [3,] 2.4283992 1.5236121 0.4836490
#>  [4,] 2.0061388 0.1710634 1.8070265
#>  [5,] 8.2909406 0.9713956 2.5100214
#>  [6,] 1.0378028 1.8871594 0.5273842
#>  [7,] 0.1126734 2.0123948 0.1707702
#>  [8,] 0.9136815 1.8168197 3.2995851
#>  [9,] 1.0147262 1.5074167 1.5125336
#> [10,] 0.5888029 1.4431174 0.1222780
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6781615 1.6781615 1.6781615
#>  [2,] 1.1909238 1.1909238 1.1909238
#>  [3,] 0.8774895 0.8774895 0.8774895
#>  [4,] 0.5892338 0.5892338 0.5892338
#>  [5,] 1.1121833 1.1121833 1.1121833
#>  [6,] 1.0063767 1.0063767 1.0063767
#>  [7,] 1.4787783 1.4787783 1.4787783
#>  [8,] 1.5819133 1.5819133 1.5819133
#>  [9,] 0.8449040 0.8449040 0.8449040
#> [10,] 0.5844634 0.5844634 0.5844634
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1475749 0.36530853 0.52116864
#>  [2,] 0.3926652 2.88552584 1.15901979
#>  [3,] 0.4916035 1.30471884 0.08906995
#>  [4,] 5.3592627 1.65147209 1.86399658
#>  [5,] 0.4715413 0.14509576 0.28380803
#>  [6,] 3.2624305 0.03947977 0.04722151
#>  [7,] 3.0585035 0.20365596 0.23610832
#>  [8,] 0.8385894 1.71236743 1.02599282
#>  [9,] 0.1370904 1.00895662 1.02849915
#> [10,] 0.1221293 2.71105245 0.58715446

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6289951 0.62899511 0.62899511
#>  [2,] 1.7818253 0.09452319 0.20456697
#>  [3,] 0.2718082 1.34043665 1.10099909
#>  [4,] 2.3136952 0.15818334 0.44773327
#>  [5,] 0.1577960 0.15779604 0.05743364
#>  [6,] 0.8536679 0.68751967 0.58817148
#>  [7,] 2.0835994 0.23967598 0.78918126
#>  [8,] 0.3385703 0.33857029 0.33857029
#>  [9,] 0.4608196 0.62625662 0.05467640
#> [10,] 0.7657719 0.89767517 0.20375600
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4984473 0.36630990 0.2485400
#>  [2,] 0.3041133 0.71344237 0.7134424
#>  [3,] 0.1381847 0.41588720 0.1381847
#>  [4,] 0.9832536 0.04000787 0.4507214
#>  [5,] 0.2155203 0.21552029 1.0845796
#>  [6,] 0.2260344 0.17663832 0.1086573
#>  [7,] 2.1328855 0.46052252 1.7050563
#>  [8,] 0.5064655 2.68507189 1.5747754
#>  [9,] 0.1594974 0.20436446 0.2043645
#> [10,] 0.1533178 0.78313415 1.3867205
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21839410 0.30169607 0.67426291
#>  [2,] 0.07942331 0.07942331 0.07942331
#>  [3,] 0.57291648 0.57291648 0.04742244
#>  [4,] 1.02924750 0.02986529 1.02924750
#>  [5,] 0.90662584 0.95677222 0.76384674
#>  [6,] 1.09562417 2.10367002 1.18118003
#>  [7,] 0.09279456 0.25213261 0.25213261
#>  [8,] 0.97126218 0.38379474 0.97126218
#>  [9,] 0.16538976 1.18471078 1.13904453
#> [10,] 1.58732852 1.08373653 0.56492800
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>            [,1]      [,2]        [,3]
#>  [1,] 0.2651473 0.2651473 0.046782258
#>  [2,] 0.3823067 0.4153087 0.382306720
#>  [3,] 0.2569958 0.2073185 0.256995769
#>  [4,] 0.1233778 0.1233778 0.123377778
#>  [5,] 0.4958258 0.4958258 2.469216116
#>  [6,] 0.6882110 0.6882110 0.003549455
#>  [7,] 0.7860598 0.6728885 0.672888482
#>  [8,] 0.3817089 0.3817089 0.843898887
#>  [9,] 0.3464249 0.3464249 0.346424898
#> [10,] 0.2002584 1.7426605 0.723702521
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.716581410 0.086179073 0.086431028
#>  [2,] 0.817914518 0.496771075 0.372345796
#>  [3,] 1.335916031 0.127696681 0.323793536
#>  [4,] 1.167535905 1.167535905 1.167535905
#>  [5,] 3.560888083 1.224855713 0.669394297
#>  [6,] 0.488126577 0.488126577 0.488126577
#>  [7,] 3.625113770 1.142933298 1.069449144
#>  [8,] 0.005826907 0.005826907 0.002044838
#>  [9,] 0.729079306 0.885093113 0.587793708
#> [10,] 0.019532650 0.019532650 0.019532650

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]        [,2]        [,3]
#>  [1,] 2.24753014 0.671941745 0.605257873
#>  [2,] 0.21171825 0.584902576 0.616648622
#>  [3,] 1.53891082 0.233346389 0.645478366
#>  [4,] 0.47274132 1.080589987 4.746740898
#>  [5,] 3.65251617 0.285001712 2.109821302
#>  [6,] 0.81757160 0.448488570 0.609669094
#>  [7,] 0.03903916 0.006238681 0.006238681
#>  [8,] 0.05152404 2.571828669 0.538949288
#>  [9,] 0.44701605 2.311067072 0.937461707
#> [10,] 2.37407783 0.052356414 2.133251433
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.5815471 1.9258307 1.6638863
#>  [2,] 0.1726112 0.9619837 0.2492228
#>  [3,] 0.2236244 0.2236244 0.2236244
#>  [4,] 1.3658827 0.3454321 2.2055242
#>  [5,] 2.0314387 0.2866543 0.6365898
#>  [6,] 3.8061227 5.2350798 3.9450112
#>  [7,] 0.5027279 0.3241752 0.3512671
#>  [8,] 4.8547983 0.4928290 0.2115034
#>  [9,] 0.3029593 0.8201953 0.9116150
#> [10,] 0.1947525 0.1947525 0.1947525
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2399714 0.2399714 0.07531588
#>  [2,] 0.6102740 0.0686966 2.33383968
#>  [3,] 0.0679248 0.0679248 0.06792480
#>  [4,] 0.4075743 7.4895158 0.84418597
#>  [5,] 1.0074658 0.6901212 0.71418070
#>  [6,] 0.3567334 1.6810900 0.34633975
#>  [7,] 1.0402816 1.5720135 0.21512419
#>  [8,] 1.8017437 0.2031325 0.40096382
#>  [9,] 1.0124791 1.9928547 0.76274912
#> [10,] 0.1672020 0.1672020 0.16720204
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]        [,2]      [,3]
#>  [1,] 3.0993173 3.635274983 3.6820275
#>  [2,] 4.0634182 0.523184832 6.0833944
#>  [3,] 1.5520122 1.552012238 1.5520122
#>  [4,] 1.8335654 1.512330825 0.2871879
#>  [5,] 1.9129853 1.442408813 2.2650904
#>  [6,] 1.6560897 0.226300198 1.6560897
#>  [7,] 0.1154436 4.937177912 0.2893114
#>  [8,] 0.2793487 1.724293565 0.6935284
#>  [9,] 3.4704978 0.003257872 0.1903132
#> [10,] 0.4987882 0.498788151 0.4987882
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.60161431 2.6290508 5.2838567
#>  [2,] 3.81744013 3.0510505 2.7301333
#>  [3,] 3.30380006 1.4774410 4.0859745
#>  [4,] 1.72701987 1.7270199 0.7805603
#>  [5,] 0.09497873 0.7550458 0.2975858
#>  [6,] 0.20933604 2.1367805 2.1367805
#>  [7,] 0.38813022 0.3881302 0.2153450
#>  [8,] 0.16848112 0.1684811 0.1684811
#>  [9,] 3.14324563 4.3186008 1.8568258
#> [10,] 1.02078289 2.0566560 2.0566560

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35050419 0.35050419 0.35050419
#>  [2,] 0.89008814 0.87323295 0.89008814
#>  [3,] 0.87307138 0.77454755 0.87307138
#>  [4,] 0.13755500 0.13755500 0.13755500
#>  [5,] 0.11462075 0.26223785 0.26223785
#>  [6,] 0.03372533 0.08891801 0.03372533
#>  [7,] 0.01209373 0.01209373 0.01209373
#>  [8,] 0.59983652 0.49599962 0.39053548
#>  [9,] 1.15258826 1.47263292 1.15258826
#> [10,] 0.17316242 0.35586232 0.35586232
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.68138512 0.37852952 0.52355503
#>  [2,] 1.46079756 1.37970860 1.49219707
#>  [3,] 0.97719607 1.05263125 1.22054589
#>  [4,] 0.27461273 0.27461273 0.48565275
#>  [5,] 0.24997617 0.24997617 0.24997617
#>  [6,] 0.61101215 0.61101215 0.25439600
#>  [7,] 0.06245059 0.22823782 0.06245059
#>  [8,] 0.02330434 0.02330434 0.02330434
#>  [9,] 0.34028672 2.05244894 2.05244894
#> [10,] 0.07962892 0.07962892 0.07962892
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.54054864 0.624504175 0.20024757
#>  [2,] 0.10443224 0.009629287 0.10443224
#>  [3,] 0.08244341 0.082443409 0.08244341
#>  [4,] 0.42410887 0.206860199 0.42410887
#>  [5,] 0.63116611 0.288660077 0.48610008
#>  [6,] 0.18179421 0.181794214 0.39948713
#>  [7,] 0.06566572 0.065665721 0.06566572
#>  [8,] 0.55296700 1.740253840 0.74773069
#>  [9,] 0.16738428 0.167384281 0.16738428
#> [10,] 0.56351454 0.563514537 0.56351454
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.73798630 0.03764129 0.73798630
#>  [2,] 1.05019313 1.05019313 0.48226613
#>  [3,] 0.31564093 0.31564093 0.31564093
#>  [4,] 0.04881029 0.04881029 0.04881029
#>  [5,] 0.38773743 0.38773743 0.38508774
#>  [6,] 1.05511727 1.05511727 1.05511727
#>  [7,] 0.11144525 0.11144525 0.11144525
#>  [8,] 0.11038211 1.30665723 0.74158242
#>  [9,] 0.11015237 0.88036837 1.18606229
#> [10,] 0.97456208 0.58290542 0.45759440
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.07917262 0.07917262 0.079172619
#>  [2,] 0.06591441 0.08834834 0.008816613
#>  [3,] 0.05835627 0.05835627 0.735351498
#>  [4,] 0.17409646 0.41822242 2.754413405
#>  [5,] 0.70758416 0.70758416 0.938051963
#>  [6,] 0.66614353 0.42085951 1.112747706
#>  [7,] 0.01506348 0.01506348 0.015063476
#>  [8,] 0.73819693 0.73819693 0.738196927
#>  [9,] 0.91302907 0.78072457 0.541078357
#> [10,] 0.03170674 0.03170674 0.031706744

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02358834 0.02358834 0.02358834
#>  [2,] 0.12565631 0.33901201 0.19541424
#>  [3,] 0.49939246 0.49939246 0.49939246
#>  [4,] 0.24077393 0.46767961 0.07846337
#>  [5,] 2.05699732 0.31481348 1.29779960
#>  [6,] 1.34525386 0.78576764 0.88925474
#>  [7,] 0.31066017 0.31066017 0.31066017
#>  [8,] 0.21862064 0.16683000 0.23881318
#>  [9,] 0.30140569 0.15202890 0.18425266
#> [10,] 0.41937731 0.27749318 0.21699079
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58002925 0.40520228 0.52694835
#>  [2,] 0.54177522 0.54177522 0.24043158
#>  [3,] 0.03238607 0.15396836 0.15396836
#>  [4,] 0.35887852 0.49471580 0.14978679
#>  [5,] 0.01911677 0.88425041 0.02811786
#>  [6,] 0.09931067 0.18716223 0.18716223
#>  [7,] 0.32643266 0.41131453 0.32643266
#>  [8,] 1.24928306 0.07853784 0.07853784
#>  [9,] 0.03758770 0.67269525 1.31437248
#> [10,] 0.11088062 0.17437099 0.11088062
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23535747 0.67247984 0.67247984
#>  [2,] 0.05830428 0.05830428 0.03104347
#>  [3,] 0.10958672 0.26713611 0.26713611
#>  [4,] 0.33820888 0.63214934 1.08969754
#>  [5,] 0.09076562 0.37584418 0.26221956
#>  [6,] 1.21111409 0.04280026 0.37570094
#>  [7,] 0.65698868 0.24657525 0.65698868
#>  [8,] 0.60052095 0.20518114 0.20518114
#>  [9,] 1.54455712 0.35148965 0.67525226
#> [10,] 0.01904154 0.52239037 0.42485441
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11208144 0.11208144 0.11208144
#>  [2,] 0.16115640 0.16115640 0.16115640
#>  [3,] 0.04731708 0.53529262 0.53529262
#>  [4,] 0.02676546 0.02676546 0.02676546
#>  [5,] 0.68326813 0.21966144 0.10591993
#>  [6,] 0.79462674 0.06305210 0.26778685
#>  [7,] 0.01063274 0.01063274 0.01063274
#>  [8,] 0.04056310 0.57932795 0.05830971
#>  [9,] 0.11494130 0.37310230 0.01539857
#> [10,] 0.04296510 0.04296510 0.04296510

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23143345 0.23143345 0.85237603
#>  [2,] 0.04868707 0.04868707 0.04868707
#>  [3,] 0.05230264 0.05230264 0.05230264
#>  [4,] 0.41833045 0.09371752 0.10266502
#>  [5,] 0.23814637 0.13171051 0.16966164
#>  [6,] 0.52949578 0.46836723 0.52341886
#>  [7,] 0.13469845 0.92675895 0.92675895
#>  [8,] 0.16020202 0.92799017 0.84249742
#>  [9,] 0.06418842 0.32911592 0.12927723
#> [10,] 0.32817812 0.51821638 0.45079116
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.99166124 1.3504191 0.50844779
#>  [2,] 0.35005028 0.6552080 1.05218610
#>  [3,] 0.11721958 0.1024972 0.45184984
#>  [4,] 1.79090101 0.4196310 1.10349870
#>  [5,] 0.76211626 0.3361043 0.48624954
#>  [6,] 0.10079386 0.6111210 0.10079386
#>  [7,] 0.68087499 1.0454464 0.55099845
#>  [8,] 0.37231488 0.4201995 0.08589404
#>  [9,] 0.06206282 0.1880395 0.16646302
#> [10,] 0.10941534 0.1821302 0.18213023
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3248603 0.07692985 0.27969635
#>  [2,] 0.1742037 0.08893765 0.26975067
#>  [3,] 0.2152185 0.21521852 0.21521852
#>  [4,] 0.1745479 0.05955757 0.04876296
#>  [5,] 0.1024449 0.75787823 1.23174439
#>  [6,] 0.6881256 0.07359350 0.07359350
#>  [7,] 0.1590729 0.15907286 0.15907286
#>  [8,] 0.1998739 0.19987389 0.19987389
#>  [9,] 0.4433886 0.04329014 0.25094181
#> [10,] 0.2234659 0.62821534 0.22346592
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28564138 0.28564138 0.21018432
#>  [2,] 0.47901602 1.02113361 1.02113361
#>  [3,] 0.18055968 0.10754797 0.15598897
#>  [4,] 0.04252139 0.02872291 0.04252139
#>  [5,] 0.25832830 0.07607810 0.52891883
#>  [6,] 0.26976463 0.03427404 0.07666398
#>  [7,] 0.88362788 0.52965041 0.92314279
#>  [8,] 0.11992745 0.04318277 0.42949149
#>  [9,] 0.27399077 0.12598050 0.02011820
#> [10,] 1.04702193 0.37049534 0.45022090

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.267629040 0.54906765 0.02093265
#>  [2,] 0.187276344 0.60538058 0.18727634
#>  [3,] 0.622670246 0.59023094 0.62267025
#>  [4,] 0.098310648 0.04610683 0.04085759
#>  [5,] 0.670223211 0.61790596 0.61790596
#>  [6,] 0.457215266 0.06628125 0.49042705
#>  [7,] 1.228219454 0.47688900 0.90717652
#>  [8,] 0.085543732 0.12616898 0.08554373
#>  [9,] 0.130375026 0.13037503 0.13037503
#> [10,] 0.009689463 0.57072290 0.88872778
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.01858958 0.13700060 0.137000603
#>  [2,] 0.01375064 0.17391955 0.285808073
#>  [3,] 0.03272541 0.03272541 0.008011742
#>  [4,] 0.02446199 0.02446199 0.024461987
#>  [5,] 0.56878391 0.11522801 0.480628798
#>  [6,] 1.08685684 0.98592613 0.218373022
#>  [7,] 1.07561775 0.22787926 0.227879263
#>  [8,] 0.13758729 0.34359952 0.137587288
#>  [9,] 0.41031039 0.99851679 0.515491910
#> [10,] 0.41984874 0.41984874 0.419848740
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.007895347 0.23772482 0.62778413
#>  [2,] 0.596030088 0.07484666 0.07484666
#>  [3,] 0.095122480 0.09512248 0.33183701
#>  [4,] 0.211154309 0.21115431 0.21115431
#>  [5,] 0.168234101 0.16823410 0.10186068
#>  [6,] 0.364459574 0.26266471 0.55788835
#>  [7,] 0.071440061 0.07144006 0.07144006
#>  [8,] 0.164874763 0.16487476 0.16487476
#>  [9,] 0.646617176 0.64661718 1.59977941
#> [10,] 0.544668934 0.43722175 0.26985672
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02238239 0.16476056 0.13087964
#>  [2,] 0.33848258 0.33848258 0.33848258
#>  [3,] 1.68946258 1.68946258 1.16437962
#>  [4,] 0.24483031 0.45429102 0.47148727
#>  [5,] 0.15172769 0.06919426 0.06919426
#>  [6,] 0.40922498 1.13874645 0.32632279
#>  [7,] 0.27485906 0.27485906 0.21994322
#>  [8,] 0.19863942 0.33518004 0.03031569
#>  [9,] 0.49697295 0.52333788 0.65628501
#> [10,] 0.64175922 0.30920627 0.58121388