Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.9659673 0.1481631 0.9957814
#>  [2,] 0.4493428 0.4493428 0.4493428
#>  [3,] 0.5254829 0.5254829 0.5254829
#>  [4,] 2.0018408 1.0450675 2.8644477
#>  [5,] 0.7459029 0.7459029 0.6675958
#>  [6,] 1.9654678 1.8159356 1.9654678
#>  [7,] 0.2583387 0.9134969 1.4485748
#>  [8,] 1.3031902 0.5685291 1.3031902
#>  [9,] 2.3962531 2.3143390 1.4563270
#> [10,] 1.2806955 0.6081489 1.6586700
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1645798 0.1645798 0.1645798
#>  [2,] 0.1143753 0.1143753 0.1143753
#>  [3,] 0.6232955 0.6232955 0.6232955
#>  [4,] 1.6970933 1.6970933 1.6970933
#>  [5,] 0.7047772 0.7047772 0.7047772
#>  [6,] 1.1485897 1.1485897 1.1485897
#>  [7,] 0.1713087 0.1713087 0.1713087
#>  [8,] 0.3720038 0.3720038 0.3720038
#>  [9,] 0.0928118 0.0928118 0.0928118
#> [10,] 1.2127110 1.2127110 1.2127110
## independence
rpextmo(10, 3, b = 1)
#>             [,1]      [,2]        [,3]
#>  [1,] 0.92960169 0.1158039 0.733269946
#>  [2,] 0.15404354 0.9523509 0.263601915
#>  [3,] 0.33954085 0.4420037 0.967031525
#>  [4,] 1.18084205 1.2431062 2.224722869
#>  [5,] 0.86178221 0.1988738 0.009015748
#>  [6,] 1.12075669 0.2939495 0.079306236
#>  [7,] 0.02206065 0.7086488 0.005191377
#>  [8,] 0.67674661 0.6554961 0.345732475
#>  [9,] 0.42426432 3.3313842 0.142782850
#> [10,] 0.18517847 0.6097670 0.204369278

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.62207530 0.17801461 0.04382999
#>  [2,] 0.92114168 0.92244115 0.87276100
#>  [3,] 0.07241059 0.09845569 0.09845569
#>  [4,] 1.41333155 6.28056455 3.00966071
#>  [5,] 1.37329297 0.15069877 1.12133487
#>  [6,] 1.79210095 6.43370908 0.18985482
#>  [7,] 0.77477702 0.74986546 0.77477702
#>  [8,] 2.16781084 3.21913842 0.67863089
#>  [9,] 0.88300468 2.37799099 2.05531631
#> [10,] 0.52225459 0.24591756 0.52225459
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02042022 0.02042022 0.02042022
#>  [2,] 1.41276387 1.41276387 1.41276387
#>  [3,] 1.28571621 1.28571621 1.28571621
#>  [4,] 0.72936468 0.72936468 0.72936468
#>  [5,] 0.37372160 0.37372160 0.37372160
#>  [6,] 1.41521539 1.41521539 1.41521539
#>  [7,] 0.59582854 0.59582854 0.59582854
#>  [8,] 0.29471469 0.29471469 0.29471469
#>  [9,] 1.91004465 1.91004465 1.91004465
#> [10,] 0.48716672 0.48716672 0.48716672
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.44511643 2.52727204 1.09135849
#>  [2,] 1.57288537 1.45280482 0.26637554
#>  [3,] 0.26199226 0.55979499 0.43518783
#>  [4,] 0.42720501 0.28310211 0.99375819
#>  [5,] 0.33199084 0.21841813 0.83243590
#>  [6,] 0.04548799 3.32946090 0.48175022
#>  [7,] 0.36664221 0.02571144 0.08222921
#>  [8,] 0.70425648 0.38263546 0.23150152
#>  [9,] 0.35963987 1.17609811 3.49833769
#> [10,] 3.99254546 0.91206940 1.23345039

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 5.8322163 5.8322163 0.6702695
#>  [2,] 4.7721052 2.7072148 2.5753832
#>  [3,] 2.4430335 1.0196450 2.5991231
#>  [4,] 1.2170808 0.7886989 1.1462016
#>  [5,] 0.1492077 0.3670015 0.9236744
#>  [6,] 2.4647544 4.4580121 0.1935453
#>  [7,] 0.3266977 0.7222919 1.8767490
#>  [8,] 0.8582322 2.2423231 3.8297268
#>  [9,] 2.6008594 1.5136144 1.5473061
#> [10,] 3.1115831 2.6711957 2.5749393
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8476634 0.8476634 0.8476634
#>  [2,] 0.5267213 0.5267213 0.5267213
#>  [3,] 0.9358638 0.9358638 0.9358638
#>  [4,] 1.1421180 1.1421180 1.1421180
#>  [5,] 0.4727257 0.4727257 0.4727257
#>  [6,] 0.2818571 0.2818571 0.2818571
#>  [7,] 0.5131067 0.5131067 0.5131067
#>  [8,] 0.8281596 0.8281596 0.8281596
#>  [9,] 2.2153130 2.2153130 2.2153130
#> [10,] 1.3232707 1.3232707 1.3232707
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1820162 0.67006750 0.4707422
#>  [2,] 0.5906493 0.01276333 2.1837684
#>  [3,] 0.4262256 0.66236912 1.0064373
#>  [4,] 3.4895185 0.76251510 1.3682577
#>  [5,] 0.9118492 0.64591092 0.3911306
#>  [6,] 1.2182527 0.03829736 0.6041602
#>  [7,] 0.7677763 0.23085391 0.8044815
#>  [8,] 0.4799057 1.22326285 0.7257993
#>  [9,] 0.4987044 0.69172472 1.2563454
#> [10,] 0.9547969 0.67619321 1.0259613

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.08893313 0.6517612 0.04799981
#>  [2,] 0.75547388 1.2048446 3.03300381
#>  [3,] 0.71628770 1.4022301 0.97678072
#>  [4,] 4.57159422 0.7867996 0.53682073
#>  [5,] 1.78733381 0.2647359 1.57049423
#>  [6,] 0.19071392 0.1907139 0.19071392
#>  [7,] 0.70667418 1.9683336 0.97838468
#>  [8,] 2.40503189 0.9860364 2.40503189
#>  [9,] 0.78597762 0.5667539 0.99598291
#> [10,] 1.77852669 4.8920525 2.81751693
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18114249 0.18114249 0.18114249
#>  [2,] 0.42875261 0.42875261 0.42875261
#>  [3,] 2.52536753 2.52536753 2.52536753
#>  [4,] 0.64144758 0.64144758 0.64144758
#>  [5,] 1.15213916 1.15213916 1.15213916
#>  [6,] 0.36359781 0.36359781 0.36359781
#>  [7,] 0.05344714 0.05344714 0.05344714
#>  [8,] 0.07718087 0.07718087 0.07718087
#>  [9,] 0.21478954 0.21478954 0.21478954
#> [10,] 0.98054038 0.98054038 0.98054038
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.37349487 2.9452274 0.17726657
#>  [2,] 0.10808767 6.7421521 0.48275415
#>  [3,] 2.06622804 1.9818852 0.79358804
#>  [4,] 1.39264753 0.5531856 0.07466634
#>  [5,] 0.27688082 5.2469342 0.20521980
#>  [6,] 0.52728948 4.4474690 0.05824045
#>  [7,] 0.09127993 0.3770944 0.86508792
#>  [8,] 0.12415575 0.2184366 0.08927130
#>  [9,] 1.20331396 1.0760021 2.45137144
#> [10,] 0.14086436 0.1997747 0.96418840

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]        [,2]       [,3]
#>  [1,] 1.7258051 0.579210977 0.08763927
#>  [2,] 1.9017372 0.193199817 2.31051071
#>  [3,] 0.8176163 2.539825725 0.66097585
#>  [4,] 0.8189295 0.230626316 0.84552519
#>  [5,] 1.0340685 1.236402975 0.36025528
#>  [6,] 0.0419580 0.005244891 4.45992787
#>  [7,] 2.8088985 2.011997143 6.25963322
#>  [8,] 0.3906303 1.277147619 0.72271739
#>  [9,] 0.9237052 2.166356766 0.56086260
#> [10,] 1.5468464 0.296726292 7.91802285
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.40213511 1.40213511 1.40213511
#>  [2,] 0.07336156 0.07336156 0.07336156
#>  [3,] 0.64503176 0.64503176 0.64503176
#>  [4,] 0.03963647 0.03963647 0.03963647
#>  [5,] 0.28933043 0.28933043 0.28933043
#>  [6,] 2.38366831 2.38366831 2.38366831
#>  [7,] 0.58575082 0.58575082 0.58575082
#>  [8,] 0.40826787 0.40826787 0.40826787
#>  [9,] 3.30918960 3.30918960 3.30918960
#> [10,] 0.46087893 0.46087893 0.46087893
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>              [,1]       [,2]      [,3]
#>  [1,] 0.407928367 1.34365013 1.4668825
#>  [2,] 1.015284747 0.79400353 0.4896717
#>  [3,] 0.331570354 2.45416438 2.7418100
#>  [4,] 0.117645843 0.05590426 1.0137086
#>  [5,] 1.869183254 1.13057661 0.5204106
#>  [6,] 0.083661970 0.20254995 1.4908062
#>  [7,] 2.537640223 0.47416938 0.4986662
#>  [8,] 0.902289385 0.61815246 0.6723237
#>  [9,] 0.008608871 0.61850788 0.2450412
#> [10,] 0.315456281 1.10621381 0.9623884

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1075540 0.09496036 1.1075540
#>  [2,] 0.4060333 2.72640293 2.0989376
#>  [3,] 1.8701025 1.49368170 1.1768310
#>  [4,] 0.6569608 0.65696077 0.6569608
#>  [5,] 2.1914022 0.75243874 0.8460286
#>  [6,] 0.9940924 0.91928344 1.1101426
#>  [7,] 0.6202821 0.36404892 0.6202821
#>  [8,] 0.0335191 0.03351910 0.0335191
#>  [9,] 0.8057389 0.98378532 2.4237033
#> [10,] 0.6276559 0.26203346 0.5791752
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3461219 0.3461219 0.3461219
#>  [2,] 2.1827594 2.1827594 2.1827594
#>  [3,] 2.0511273 2.0511273 2.0511273
#>  [4,] 0.1751565 0.1751565 0.1751565
#>  [5,] 0.9402940 0.9402940 0.9402940
#>  [6,] 1.2697923 1.2697923 1.2697923
#>  [7,] 1.6453052 1.6453052 1.6453052
#>  [8,] 0.1306147 0.1306147 0.1306147
#>  [9,] 0.5725860 0.5725860 0.5725860
#> [10,] 0.4693367 0.4693367 0.4693367
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.06291699 0.313432669 0.04886380
#>  [2,] 2.55259156 0.024788217 0.90771191
#>  [3,] 0.11632207 0.163114902 0.97973125
#>  [4,] 0.25288515 1.159271608 0.20355377
#>  [5,] 0.41251258 3.248325281 1.01554872
#>  [6,] 0.27078441 0.004203035 0.74542174
#>  [7,] 0.50951910 2.086665158 1.40440609
#>  [8,] 0.01162480 0.483680077 1.86559410
#>  [9,] 0.88181368 0.256014297 1.05890074
#> [10,] 0.21629175 3.387105629 0.01482523

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3541036 0.3541036 0.35410359
#>  [2,] 0.9173829 0.5772232 0.14708916
#>  [3,] 1.5976231 1.3372944 1.59762311
#>  [4,] 0.2590070 6.2533652 5.52457570
#>  [5,] 0.9804460 0.1709631 3.72696617
#>  [6,] 1.3083992 1.2371056 0.07536387
#>  [7,] 2.1602136 3.3044667 0.74567233
#>  [8,] 2.4654305 4.9604090 0.66567814
#>  [9,] 1.4655521 1.4655521 1.46555215
#> [10,] 1.4835068 3.9565741 1.67542824
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4490361 1.4490361 1.4490361
#>  [2,] 0.6421700 0.6421700 0.6421700
#>  [3,] 0.3995285 0.3995285 0.3995285
#>  [4,] 3.8397812 3.8397812 3.8397812
#>  [5,] 2.0611291 2.0611291 2.0611291
#>  [6,] 1.4977875 1.4977875 1.4977875
#>  [7,] 0.4649173 0.4649173 0.4649173
#>  [8,] 0.3033751 0.3033751 0.3033751
#>  [9,] 0.4350564 0.4350564 0.4350564
#> [10,] 0.1547929 0.1547929 0.1547929
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>            [,1]       [,2]        [,3]
#>  [1,] 3.2566043 0.39121128 1.217141638
#>  [2,] 0.8054756 0.94923034 1.544874174
#>  [3,] 0.1567742 0.20262578 0.001314823
#>  [4,] 1.7935803 4.22061389 0.130818820
#>  [5,] 1.6367050 0.51770796 3.203841588
#>  [6,] 1.4433192 0.67528278 0.368854765
#>  [7,] 4.7433571 0.06556117 0.336473269
#>  [8,] 0.4503392 0.39334327 0.416218717
#>  [9,] 1.7365359 0.19569372 0.616466062
#> [10,] 0.4321186 0.91901078 1.028752322

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 4.22837786 0.16102245 1.3291134
#>  [2,] 4.44649122 0.05959090 4.4464912
#>  [3,] 0.38554746 1.17383199 0.3071033
#>  [4,] 2.26028088 2.26028088 1.5646130
#>  [5,] 1.41076057 1.95037656 0.5368962
#>  [6,] 0.65428985 1.47472025 0.1167466
#>  [7,] 1.62043473 0.25468017 2.2641053
#>  [8,] 2.26317147 0.03351673 1.1154434
#>  [9,] 3.34020881 1.85660783 3.3402088
#> [10,] 0.05744664 0.22574441 2.9268763
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3590969 0.3590969 0.3590969
#>  [2,] 0.1491250 0.1491250 0.1491250
#>  [3,] 0.7736837 0.7736837 0.7736837
#>  [4,] 0.1562596 0.1562596 0.1562596
#>  [5,] 0.2168160 0.2168160 0.2168160
#>  [6,] 0.6093503 0.6093503 0.6093503
#>  [7,] 2.6246147 2.6246147 2.6246147
#>  [8,] 0.3089707 0.3089707 0.3089707
#>  [9,] 1.7852660 1.7852660 1.7852660
#> [10,] 3.2797466 3.2797466 3.2797466
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.35435092 0.2617690 0.23509920
#>  [2,] 0.01074282 0.2265381 2.22959652
#>  [3,] 0.28675979 1.7324334 1.97256366
#>  [4,] 1.23042504 2.6421025 0.04574542
#>  [5,] 0.01180818 0.8505721 0.41350460
#>  [6,] 0.85785657 0.2200040 0.32438419
#>  [7,] 1.65145928 0.7240548 0.55860318
#>  [8,] 0.42737103 0.0122801 0.77040145
#>  [9,] 0.35576391 1.3232761 1.94527256
#> [10,] 2.81870003 3.3585611 0.56232446

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.54755935 3.09891518 1.49300247
#>  [2,] 1.84600473 0.17776814 3.03413509
#>  [3,] 0.03875956 0.03875956 0.03875956
#>  [4,] 0.71677166 1.13481093 0.56195820
#>  [5,] 1.17294758 1.17224671 1.47434740
#>  [6,] 1.71846303 0.36203115 3.34404205
#>  [7,] 1.69895282 2.38343956 1.35991706
#>  [8,] 1.29070095 0.97367963 0.44280212
#>  [9,] 0.17746352 3.25986830 0.46874553
#> [10,] 0.04385779 0.04385779 0.04385779
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38911934 0.38911934 0.38911934
#>  [2,] 1.63266311 1.63266311 1.63266311
#>  [3,] 0.75556732 0.75556732 0.75556732
#>  [4,] 1.35146576 1.35146576 1.35146576
#>  [5,] 0.25890963 0.25890963 0.25890963
#>  [6,] 0.08989788 0.08989788 0.08989788
#>  [7,] 0.89637260 0.89637260 0.89637260
#>  [8,] 2.07912703 2.07912703 2.07912703
#>  [9,] 0.97640373 0.97640373 0.97640373
#> [10,] 0.92953617 0.92953617 0.92953617
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6087842 1.20121259 0.97856657
#>  [2,] 0.2088290 2.80879192 0.18810545
#>  [3,] 2.7306667 1.91454484 0.64608227
#>  [4,] 0.6506473 3.03760479 1.93480400
#>  [5,] 0.6740710 0.87991698 0.43231252
#>  [6,] 0.3566311 0.08931822 0.16090153
#>  [7,] 0.4004571 0.97066024 0.49726156
#>  [8,] 0.3604473 0.38097790 0.79261853
#>  [9,] 0.1496441 0.68944252 0.05750825
#> [10,] 5.0280394 0.21235029 0.31472035

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0968157 0.6930961 0.7375576
#>  [2,] 2.0196995 0.6267125 0.1782223
#>  [3,] 0.1326264 0.9134572 1.6785998
#>  [4,] 2.3702664 3.5750138 4.7617637
#>  [5,] 0.9404343 2.6529632 0.2274691
#>  [6,] 0.8400713 0.2025130 0.1669930
#>  [7,] 0.4192192 0.4192192 0.4192192
#>  [8,] 1.6882853 1.1053301 1.5667800
#>  [9,] 0.0694022 1.2205163 2.2317424
#> [10,] 0.2091159 1.8549832 3.2099168
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.3456661 2.3456661 2.3456661
#>  [2,] 0.1531825 0.1531825 0.1531825
#>  [3,] 1.3962784 1.3962784 1.3962784
#>  [4,] 0.1668193 0.1668193 0.1668193
#>  [5,] 0.1896440 0.1896440 0.1896440
#>  [6,] 1.0511711 1.0511711 1.0511711
#>  [7,] 0.8666958 0.8666958 0.8666958
#>  [8,] 2.9951691 2.9951691 2.9951691
#>  [9,] 1.6191932 1.6191932 1.6191932
#> [10,] 0.2883890 0.2883890 0.2883890
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.057174060 1.19970385 1.53310504
#>  [2,] 1.000717334 0.57624714 0.52862801
#>  [3,] 0.961707411 0.27007444 1.13598004
#>  [4,] 0.145478646 0.38117090 0.66947909
#>  [5,] 0.021675548 0.13851976 1.86756654
#>  [6,] 0.418903074 0.96994683 0.07196763
#>  [7,] 0.005226153 0.59814107 0.50760912
#>  [8,] 0.801745972 1.38912763 0.26526066
#>  [9,] 0.696790968 0.04039027 2.34040454
#> [10,] 0.607208207 0.54678693 0.02192356

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.01053413 1.314931712 1.25759529
#>  [2,] 0.14287154 0.050969355 0.14287154
#>  [3,] 0.19591272 0.288727125 0.28872712
#>  [4,] 0.90236676 0.439096664 1.04490397
#>  [5,] 0.85319169 0.959090496 0.85319169
#>  [6,] 3.12576557 0.083651021 0.08365102
#>  [7,] 0.50707529 0.238004385 0.05259950
#>  [8,] 0.19757847 0.002951297 0.70319247
#>  [9,] 0.28871510 0.288715096 1.11616456
#> [10,] 0.46321485 0.656023654 0.46321485
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7216336 0.01645568 0.1026686
#>  [2,] 1.0649119 0.96047304 1.5835947
#>  [3,] 0.1009513 1.90795483 0.7510663
#>  [4,] 0.5937594 1.79725502 0.9075701
#>  [5,] 0.0531644 1.53613962 0.0531644
#>  [6,] 1.4427379 1.44273789 0.2255399
#>  [7,] 1.5002035 0.42790762 0.8772664
#>  [8,] 0.1511814 1.65514010 1.4642334
#>  [9,] 1.3055044 1.40511301 1.0576023
#> [10,] 0.1571920 0.15719195 0.2155216
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.44387976 2.64285098 2.64285098
#>  [2,] 0.96379337 0.13883778 0.13883778
#>  [3,] 0.32474321 0.29709472 1.04380325
#>  [4,] 1.45402045 1.45402045 1.45402045
#>  [5,] 0.17434716 0.17434716 1.07350948
#>  [6,] 0.70982476 0.70982476 1.06109915
#>  [7,] 1.31262401 0.48214898 2.29626221
#>  [8,] 0.50197136 0.82901776 0.12079062
#>  [9,] 0.01755853 0.01755853 0.01755853
#> [10,] 1.68183282 0.51553448 0.24277920
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 1.05223802 0.899632976 1.63467793
#>  [2,] 0.09570732 0.762799316 0.37684714
#>  [3,] 0.13610727 0.009572929 0.48813990
#>  [4,] 0.14040623 0.441167649 0.14040623
#>  [5,] 0.80718709 0.763167607 0.16696124
#>  [6,] 1.04412268 0.073475328 0.07347533
#>  [7,] 0.10035042 0.100350421 0.10035042
#>  [8,] 0.18102619 0.445854527 0.32257593
#>  [9,] 0.49629727 0.496297274 0.14748564
#> [10,] 1.75561767 0.660843826 0.66084383
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21559833 0.41822054 0.21559833
#>  [2,] 0.39890178 0.05422824 0.05422824
#>  [3,] 1.05891443 1.27857633 1.68823722
#>  [4,] 0.73126504 0.01569217 0.07036397
#>  [5,] 0.50083963 0.21638660 0.50083963
#>  [6,] 0.16878114 0.12592720 0.12592720
#>  [7,] 1.79907508 0.74205422 1.22903059
#>  [8,] 3.44503854 1.51101978 2.45418342
#>  [9,] 1.02511584 1.96720227 0.01705872
#> [10,] 0.06341505 0.06341505 0.06341505

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]        [,3]
#>  [1,] 1.2658489 0.5509793 2.039965874
#>  [2,] 2.1925525 0.1747686 1.037295615
#>  [3,] 0.3814323 0.4222333 0.305869325
#>  [4,] 1.4598167 1.4598167 1.459816653
#>  [5,] 0.1640402 1.0576160 0.493948424
#>  [6,] 0.2134809 1.3802241 1.485017088
#>  [7,] 1.1703015 0.2957430 1.170301486
#>  [8,] 0.7204480 0.7998913 0.954474096
#>  [9,] 0.5670864 0.5670864 0.202499718
#> [10,] 0.4787035 0.5177980 0.007991522
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.97589159 3.0410218 5.3608803
#>  [2,] 0.47599655 0.4759965 0.2016358
#>  [3,] 0.27050416 0.4477195 0.5655897
#>  [4,] 0.08034869 0.5667706 0.9791892
#>  [5,] 0.27989937 0.2798994 0.2798994
#>  [6,] 0.02906810 1.1852818 4.4254944
#>  [7,] 0.85759529 0.2579488 2.3928321
#>  [8,] 0.71137383 0.7113738 0.7113738
#>  [9,] 2.08460303 0.6188800 1.3029018
#> [10,] 1.66165321 1.6616532 0.6012708
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.34633904 0.88731963 1.30690012
#>  [2,] 1.05360248 1.64957735 0.51793082
#>  [3,] 1.02304518 0.37530112 1.02304518
#>  [4,] 0.15909911 0.15909911 0.15909911
#>  [5,] 1.55368068 3.68060668 1.75433618
#>  [6,] 1.89123276 1.37948089 5.25478456
#>  [7,] 1.17665018 1.17665018 0.16399773
#>  [8,] 0.45606046 3.43650668 3.80988802
#>  [9,] 0.05690417 0.05690417 0.05690417
#> [10,] 0.07244391 1.08810612 2.97258863
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>            [,1]         [,2]       [,3]
#>  [1,] 2.9013323 0.8167041636 0.93463832
#>  [2,] 1.1730046 0.7491437612 1.17300463
#>  [3,] 3.4824424 3.4824423913 1.19124892
#>  [4,] 0.3290491 0.3290491363 0.32904914
#>  [5,] 3.4270735 0.5565802051 0.07057659
#>  [6,] 0.5467890 0.0007196718 0.54678900
#>  [7,] 1.9203802 1.9203801869 0.69824986
#>  [8,] 3.4683090 0.7628592972 0.48973878
#>  [9,] 1.4749855 1.4749854664 0.53092863
#> [10,] 2.9092476 1.6001476635 1.58434842
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.04234911 1.2756642 0.1697455
#>  [2,] 1.60651809 0.7056999 1.6065181
#>  [3,] 0.33462615 0.6228828 0.6228828
#>  [4,] 0.53014030 0.5569614 0.5569614
#>  [5,] 0.86705475 0.5890538 1.4335374
#>  [6,] 1.98750257 1.9875026 1.9875026
#>  [7,] 4.73148906 5.5104824 0.2509502
#>  [8,] 0.60804596 0.7579891 0.7579891
#>  [9,] 1.96652213 0.4268109 0.9834549
#> [10,] 5.41987848 2.2729244 3.8369606

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38157347 0.45180855 0.45180855
#>  [2,] 0.39962321 0.39962321 0.02045501
#>  [3,] 0.62021284 1.14449995 0.68776434
#>  [4,] 0.23364311 0.23364311 0.23364311
#>  [5,] 1.26225038 0.66983412 0.17955918
#>  [6,] 0.21247295 0.09023486 0.07760733
#>  [7,] 0.55765423 1.40617478 0.43537851
#>  [8,] 0.17060518 0.63818901 0.17060518
#>  [9,] 0.08811882 0.08890506 0.08890506
#> [10,] 0.18971767 0.18971767 0.18971767
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>              [,1]        [,2]         [,3]
#>  [1,] 0.131435451 0.054410230 0.2922288467
#>  [2,] 0.008071145 0.008071145 0.3338802358
#>  [3,] 0.610526505 0.380844054 0.0005512675
#>  [4,] 0.859624287 0.380367764 0.8596242870
#>  [5,] 0.196104335 0.145268157 0.1961043354
#>  [6,] 0.194935262 0.194935262 0.1949352624
#>  [7,] 1.437347779 0.849182234 1.2618139637
#>  [8,] 0.491446513 3.115110358 0.6785926385
#>  [9,] 0.135902929 0.135902929 0.1359029291
#> [10,] 0.372791755 0.306105293 3.3907684374
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5070151 0.1265776 0.03371603
#>  [2,] 0.1246707 0.1246707 0.12467075
#>  [3,] 0.3424979 0.3424979 0.34249786
#>  [4,] 0.9294013 0.8591754 0.92940127
#>  [5,] 0.3801195 0.1258778 0.09019589
#>  [6,] 1.0380615 0.7140627 1.03806146
#>  [7,] 0.0663195 0.0663195 0.06631950
#>  [8,] 0.3878648 0.3878648 0.38786477
#>  [9,] 0.4281670 0.7355399 0.73553987
#> [10,] 1.1025160 0.2695450 1.07122669
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.19828433 0.29969209 1.31150774
#>  [2,] 0.22023434 0.22023434 0.22023434
#>  [3,] 1.74092691 0.14026488 0.70295789
#>  [4,] 0.40357429 0.07646954 0.33695465
#>  [5,] 0.34748376 0.62776386 0.34748376
#>  [6,] 0.02942521 0.32166208 0.35283001
#>  [7,] 0.50465178 1.16404776 0.39234491
#>  [8,] 0.04882134 0.04882134 0.04882134
#>  [9,] 1.29672948 1.29672948 1.29672948
#> [10,] 0.02110157 0.21726181 0.21726181
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 1.350531764 1.764869868 0.480350118
#>  [2,] 0.002512019 0.002512019 0.002512019
#>  [3,] 0.970002256 0.970002256 0.434896949
#>  [4,] 1.546388878 0.558299850 0.170274018
#>  [5,] 0.602246004 0.602246004 0.602246004
#>  [6,] 0.727066823 0.727066823 0.727066823
#>  [7,] 0.650406472 0.210336365 0.692687638
#>  [8,] 0.563860425 0.654011650 0.393292939
#>  [9,] 1.824315668 1.824315668 0.172447781
#> [10,] 0.757448113 0.504557233 1.442110028

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08714075 1.16680255 1.02591703
#>  [2,] 0.12219511 0.12219511 0.03665529
#>  [3,] 0.62939322 0.98518226 0.93503962
#>  [4,] 0.06229511 0.22163057 0.22163057
#>  [5,] 1.01463991 1.03437898 0.66741476
#>  [6,] 0.57944948 1.15282913 0.90504251
#>  [7,] 0.29942467 0.28039579 0.28039579
#>  [8,] 0.75595891 0.15413593 0.47776318
#>  [9,] 0.31906037 0.63830120 0.53525770
#> [10,] 0.08246937 0.08246937 0.08246937
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08243213 0.08243213 0.08243213
#>  [2,] 0.37627364 0.37627364 0.02307547
#>  [3,] 0.72905681 0.72905681 0.72905681
#>  [4,] 0.18047078 0.18047078 0.03283050
#>  [5,] 1.42616961 1.42616961 0.76748992
#>  [6,] 0.24437685 0.28189746 0.76151956
#>  [7,] 0.13147465 0.13147465 0.08614476
#>  [8,] 0.29614370 0.18970161 0.29614370
#>  [9,] 0.02297108 0.26086178 0.25509395
#> [10,] 0.27513853 0.03894870 0.30098942
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>              [,1]         [,2]       [,3]
#>  [1,] 0.072785906 0.5573922727 0.79370543
#>  [2,] 0.549197539 0.0531903694 0.33834334
#>  [3,] 0.005296384 0.0565836764 0.80985598
#>  [4,] 0.995875612 0.0463966437 0.39160430
#>  [5,] 0.475961662 0.4759616624 0.08109680
#>  [6,] 0.622776822 0.2039764076 0.09472709
#>  [7,] 0.528290266 0.0008848545 0.75405586
#>  [8,] 0.025771673 0.0257716734 0.02577167
#>  [9,] 0.157059641 0.1570596408 0.15705964
#> [10,] 0.103561068 0.4266019769 0.42660198
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.648755156 1.60850655 1.06697469
#>  [2,] 0.335867480 0.19923405 0.59557891
#>  [3,] 0.133293754 0.05906721 0.13329375
#>  [4,] 1.974305244 0.44393783 0.52780053
#>  [5,] 0.537978104 1.54648830 0.70939561
#>  [6,] 0.007760482 0.51359475 0.51359475
#>  [7,] 0.936634815 0.35108851 0.35108851
#>  [8,] 0.362000531 0.44237183 0.07221554
#>  [9,] 0.247921590 0.24792159 1.02629139
#> [10,] 0.205652703 0.63579162 0.07043483

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.97245837 0.04055492 0.25317498
#>  [2,] 0.39805455 0.13772129 0.55681215
#>  [3,] 0.33853055 0.46666988 0.59836822
#>  [4,] 0.03263073 0.34070857 0.13483054
#>  [5,] 0.25230399 0.09581102 0.09581102
#>  [6,] 0.56968847 0.56968847 1.33872701
#>  [7,] 0.05481980 0.74690034 0.09892335
#>  [8,] 0.76868344 0.04706309 0.04706309
#>  [9,] 0.00612082 0.19368382 0.29023167
#> [10,] 0.07406743 0.36345358 0.83671770
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.344315721 1.225512524 1.225512524
#>  [2,] 0.156213110 0.156213110 0.156213110
#>  [3,] 0.628111940 0.557514387 0.300248756
#>  [4,] 0.552015237 0.058157248 0.552015237
#>  [5,] 0.037397073 0.037397073 0.913341112
#>  [6,] 1.206673863 0.392229834 0.491013131
#>  [7,] 0.005943762 0.005943762 0.005943762
#>  [8,] 0.182796186 0.280098627 0.280098627
#>  [9,] 0.092546236 0.092546236 0.092546236
#> [10,] 0.107744852 0.486954530 0.085086024
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.68048920 0.452456677 0.830181551
#>  [2,] 0.36999353 0.197548912 1.016557861
#>  [3,] 0.15559252 0.799202414 0.175312000
#>  [4,] 0.38717299 1.282202979 1.282202979
#>  [5,] 0.05453766 0.003634358 0.493512721
#>  [6,] 1.17737517 0.068857086 0.048721953
#>  [7,] 0.09278913 0.092789130 0.092789130
#>  [8,] 2.29082279 0.375522496 0.008212759
#>  [9,] 1.33309269 0.434676808 0.540292007
#> [10,] 0.02779768 0.198657790 0.198657790
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.29414292 1.51268699 1.123883728
#>  [2,] 0.46542498 0.38969608 0.465424980
#>  [3,] 0.49692038 0.86928208 0.982880810
#>  [4,] 0.02228897 0.02228897 0.022288975
#>  [5,] 0.10968536 0.10968536 0.447005606
#>  [6,] 0.44195815 0.31297638 0.001977563
#>  [7,] 0.33767700 0.39463449 0.532795839
#>  [8,] 2.23647406 0.51166425 0.877876827
#>  [9,] 0.22595890 0.64219638 1.026105202
#> [10,] 0.15453996 0.10717847 0.078460600

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.05672690 0.31713742 0.3171374
#>  [2,] 0.21320161 0.93529985 0.8387262
#>  [3,] 0.97305094 0.93684828 0.2462010
#>  [4,] 0.19948230 0.19948230 0.1994823
#>  [5,] 0.15961337 0.90355140 0.1596134
#>  [6,] 0.37078552 0.45382229 0.8692545
#>  [7,] 0.31689677 0.08368158 0.3168968
#>  [8,] 0.10629657 0.10629657 0.1062966
#>  [9,] 0.70659326 2.01914191 0.7065933
#> [10,] 0.06906612 0.08855167 0.4668354
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23139703 0.23139703 0.14663788
#>  [2,] 2.05124761 0.43689080 0.38388643
#>  [3,] 0.60015740 0.60015740 0.25891096
#>  [4,] 0.01175052 0.01175052 0.01175052
#>  [5,] 0.10734503 0.10734503 0.08193545
#>  [6,] 0.75586240 0.22869568 0.14407916
#>  [7,] 0.53338079 0.01410497 0.53338079
#>  [8,] 0.39069508 0.05079489 1.65090267
#>  [9,] 0.06975793 0.06975793 0.06975793
#> [10,] 0.06334476 0.37050567 0.11937125
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.32579844 0.554463076 0.29862740
#>  [2,] 0.05191765 0.051568258 0.05191765
#>  [3,] 0.39364659 0.393646587 0.39364659
#>  [4,] 0.87963363 0.339363641 0.06956173
#>  [5,] 0.06018641 0.060186406 0.06018641
#>  [6,] 0.07463989 0.074639887 0.04353235
#>  [7,] 0.02214944 0.005493169 0.09766242
#>  [8,] 0.82806754 0.654660483 1.03651097
#>  [9,] 0.97528028 0.195035957 0.10352858
#> [10,] 0.73887525 0.125533561 0.73490260
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.69429675 0.69429675 0.48894854
#>  [2,] 0.12234914 0.12234914 1.57154305
#>  [3,] 0.08652128 0.15716791 0.08652128
#>  [4,] 0.03383553 0.03383553 0.11528284
#>  [5,] 0.01881149 0.01881149 0.01881149
#>  [6,] 0.05722981 0.05722981 0.05722981
#>  [7,] 1.06428965 0.57411439 1.82296842
#>  [8,] 0.08809995 0.20351555 0.17014548
#>  [9,] 0.34298475 0.34298475 0.34298475
#> [10,] 0.03646205 0.03646205 0.76239853