Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2515643 1.4930686 0.03387426
#>  [2,] 0.1159625 0.9454778 0.38628404
#>  [3,] 0.6201419 0.5889580 1.21719914
#>  [4,] 1.1478544 2.3649288 1.78332808
#>  [5,] 0.4260703 0.3654087 1.18976875
#>  [6,] 1.0474724 0.2529175 3.98753559
#>  [7,] 0.0445466 0.2565447 0.25654475
#>  [8,] 1.0739298 1.0739298 0.33681779
#>  [9,] 0.3264424 0.2381815 1.36844347
#> [10,] 0.2476888 0.1162603 5.00584565
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0906891 1.0906891 1.0906891
#>  [2,] 1.0325644 1.0325644 1.0325644
#>  [3,] 0.7515231 0.7515231 0.7515231
#>  [4,] 0.4789534 0.4789534 0.4789534
#>  [5,] 0.3768095 0.3768095 0.3768095
#>  [6,] 1.6900358 1.6900358 1.6900358
#>  [7,] 0.8930080 0.8930080 0.8930080
#>  [8,] 0.9629638 0.9629638 0.9629638
#>  [9,] 0.2453832 0.2453832 0.2453832
#> [10,] 2.1753256 2.1753256 2.1753256
## independence
rpextmo(10, 3, b = 1)
#>             [,1]      [,2]       [,3]
#>  [1,] 2.91569567 1.4177656 0.82018492
#>  [2,] 0.02859253 2.3989849 0.31930530
#>  [3,] 0.42300168 0.6246209 0.46281716
#>  [4,] 0.69328839 0.3027057 0.57247755
#>  [5,] 0.61487338 0.7716601 0.17748622
#>  [6,] 2.31037513 1.4982602 2.67747890
#>  [7,] 0.60050716 0.7331599 0.36377191
#>  [8,] 0.50297561 1.1050098 0.73249969
#>  [9,] 0.18872667 1.3482385 1.03375980
#> [10,] 0.28301877 0.4076435 0.02792947

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.1779243 1.1779243 1.17792430
#>  [2,] 0.3098396 0.3098396 0.30983960
#>  [3,] 0.4181793 0.3085808 2.98425806
#>  [4,] 4.9458837 0.3794551 3.10763731
#>  [5,] 0.9867589 0.4621583 0.60106905
#>  [6,] 1.2574376 1.2574376 1.25743760
#>  [7,] 1.7906616 2.6168890 1.35369568
#>  [8,] 0.1255905 3.5324360 0.38127554
#>  [9,] 2.5971078 0.5498365 1.29069097
#> [10,] 2.4361698 2.6081278 0.04550162
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.56160779 1.56160779 1.56160779
#>  [2,] 3.33768831 3.33768831 3.33768831
#>  [3,] 0.31977988 0.31977988 0.31977988
#>  [4,] 1.20004439 1.20004439 1.20004439
#>  [5,] 0.03177221 0.03177221 0.03177221
#>  [6,] 3.52156167 3.52156167 3.52156167
#>  [7,] 0.92162182 0.92162182 0.92162182
#>  [8,] 0.02421567 0.02421567 0.02421567
#>  [9,] 0.26319078 0.26319078 0.26319078
#> [10,] 0.11511049 0.11511049 0.11511049
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1115469 0.2514686 0.72410339
#>  [2,] 0.3732987 0.8393646 0.63155507
#>  [3,] 0.3141672 0.2302183 0.10706420
#>  [4,] 0.1582955 1.6391137 0.72324925
#>  [5,] 0.9968725 0.7523508 1.08033762
#>  [6,] 0.4810291 3.4358849 0.05798025
#>  [7,] 0.3849651 0.6125459 0.24757400
#>  [8,] 1.5229957 1.5965846 0.66852866
#>  [9,] 1.3589312 1.7950596 0.27793038
#> [10,] 0.5073609 1.4470443 1.02331433

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.6673466 1.4684100 1.66734661
#>  [2,] 0.8555669 1.3524076 1.35240760
#>  [3,] 2.4046340 0.4417973 1.62650174
#>  [4,] 0.2267786 0.2267786 0.22677857
#>  [5,] 0.1768028 0.1579318 0.04945326
#>  [6,] 0.1432599 1.1908767 1.67462303
#>  [7,] 1.4463885 0.7658327 0.56729611
#>  [8,] 2.8766066 1.1622983 2.25527923
#>  [9,] 1.7292557 3.2174282 2.18631477
#> [10,] 1.8307632 0.2249829 2.75667899
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07561596 0.07561596 0.07561596
#>  [2,] 1.15797265 1.15797265 1.15797265
#>  [3,] 0.01099923 0.01099923 0.01099923
#>  [4,] 0.87255014 0.87255014 0.87255014
#>  [5,] 0.35095649 0.35095649 0.35095649
#>  [6,] 0.37929650 0.37929650 0.37929650
#>  [7,] 0.78836666 0.78836666 0.78836666
#>  [8,] 1.24519294 1.24519294 1.24519294
#>  [9,] 2.59187583 2.59187583 2.59187583
#> [10,] 0.81870339 0.81870339 0.81870339
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0248500 0.82275914 0.32239066
#>  [2,] 1.0644616 0.46306271 1.51202539
#>  [3,] 1.8884658 0.77079673 0.85638612
#>  [4,] 0.5158595 0.64756599 0.04962579
#>  [5,] 0.6997181 0.35787408 0.31056800
#>  [6,] 1.1041725 1.80677035 0.44590981
#>  [7,] 0.5296966 0.06196983 0.08461468
#>  [8,] 1.6863346 0.07608132 5.37532751
#>  [9,] 4.1260949 0.24268978 0.71582612
#> [10,] 2.2751008 1.86862585 0.96691792

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.7582993 4.5165440 0.32648384
#>  [2,] 0.4436627 1.2151592 1.05891194
#>  [3,] 2.1588849 0.4244524 0.87626117
#>  [4,] 0.2580790 0.5620892 0.56208924
#>  [5,] 2.5188122 1.4997553 1.30727782
#>  [6,] 0.3564811 0.2311501 0.35648111
#>  [7,] 0.9769598 3.3937745 0.71381361
#>  [8,] 0.2304503 0.5225338 0.52593674
#>  [9,] 0.6195693 0.6195693 0.61956929
#> [10,] 1.0977168 0.3032307 0.05732088
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.000898669 0.000898669 0.000898669
#>  [2,] 0.183699307 0.183699307 0.183699307
#>  [3,] 1.627893273 1.627893273 1.627893273
#>  [4,] 0.586102411 0.586102411 0.586102411
#>  [5,] 0.416067218 0.416067218 0.416067218
#>  [6,] 0.967388885 0.967388885 0.967388885
#>  [7,] 0.049187897 0.049187897 0.049187897
#>  [8,] 0.113560966 0.113560966 0.113560966
#>  [9,] 0.728165600 0.728165600 0.728165600
#> [10,] 0.190992692 0.190992692 0.190992692
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1210994 2.80345422 2.3714157
#>  [2,] 0.9383161 1.54895570 0.4858772
#>  [3,] 2.1011136 1.07635441 1.4734529
#>  [4,] 0.8213617 2.58507926 0.6940082
#>  [5,] 0.7179893 1.73268581 3.3514379
#>  [6,] 0.2571045 0.03818379 0.2478782
#>  [7,] 0.8040135 4.15015956 1.4422790
#>  [8,] 0.4464988 0.32210853 1.2795375
#>  [9,] 1.2139222 0.40687006 1.9881411
#> [10,] 0.6803863 4.68723306 0.3734324

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3445682 0.2445543 1.1369561
#>  [2,] 0.3260603 0.3260603 0.3260603
#>  [3,] 0.5576990 3.9038761 1.1771572
#>  [4,] 0.6077709 0.6077709 0.6077709
#>  [5,] 0.5698599 1.0133725 2.1678741
#>  [6,] 0.6287838 0.5752184 0.6287838
#>  [7,] 0.9942690 0.9942690 0.2456745
#>  [8,] 0.3535864 0.2940432 4.8319526
#>  [9,] 9.6467001 2.6312124 1.2637617
#> [10,] 0.7747190 4.7665987 0.5541019
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.41178478 1.41178478 1.41178478
#>  [2,] 1.61086611 1.61086611 1.61086611
#>  [3,] 0.94881961 0.94881961 0.94881961
#>  [4,] 0.14013577 0.14013577 0.14013577
#>  [5,] 1.61514707 1.61514707 1.61514707
#>  [6,] 0.41699511 0.41699511 0.41699511
#>  [7,] 0.08524759 0.08524759 0.08524759
#>  [8,] 2.36016100 2.36016100 2.36016100
#>  [9,] 0.07526534 0.07526534 0.07526534
#> [10,] 1.25951619 1.25951619 1.25951619
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4554188 1.3270504 0.19393768
#>  [2,] 1.1282807 0.7863853 1.44798141
#>  [3,] 0.2825205 1.7616605 2.40534968
#>  [4,] 1.0815299 0.1859396 0.76137332
#>  [5,] 0.8695540 0.2607622 1.02058950
#>  [6,] 0.5972557 0.7096730 0.82483015
#>  [7,] 1.5381440 0.7869344 1.84742135
#>  [8,] 1.1999170 1.5913866 0.08463684
#>  [9,] 2.6238114 1.4910067 1.74077831
#> [10,] 1.1816722 0.5898435 1.10745829

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]       [,2]      [,3]
#>  [1,] 3.09885507 3.97699735 3.2564641
#>  [2,] 0.49095145 0.06984764 1.1057315
#>  [3,] 0.87594233 0.24840591 4.6696862
#>  [4,] 0.10157596 0.71855586 0.7237697
#>  [5,] 0.37573813 0.37573813 0.3757381
#>  [6,] 1.06878242 1.22689628 0.4266073
#>  [7,] 0.56980923 2.12620317 1.7173784
#>  [8,] 2.83167887 1.46460660 0.8286074
#>  [9,] 0.07479984 1.45891200 1.4589120
#> [10,] 0.50662172 0.50662172 0.5066217
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1805224 0.1805224 0.1805224
#>  [2,] 1.4152074 1.4152074 1.4152074
#>  [3,] 0.7175071 0.7175071 0.7175071
#>  [4,] 1.7890518 1.7890518 1.7890518
#>  [5,] 0.2931137 0.2931137 0.2931137
#>  [6,] 0.2227488 0.2227488 0.2227488
#>  [7,] 0.1601761 0.1601761 0.1601761
#>  [8,] 0.5017943 0.5017943 0.5017943
#>  [9,] 1.6774291 1.6774291 1.6774291
#> [10,] 1.4738576 1.4738576 1.4738576
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.35771443 2.93169816 0.4618680
#>  [2,] 0.25620614 0.74636225 0.4756702
#>  [3,] 0.05030715 2.68165307 0.1833523
#>  [4,] 0.32130949 2.05872415 0.8713682
#>  [5,] 0.70401932 0.52774549 0.2003997
#>  [6,] 0.66652607 2.27211692 0.4827728
#>  [7,] 0.11106994 0.61269820 0.6191929
#>  [8,] 0.44129232 0.45993324 0.7345660
#>  [9,] 0.67584810 0.07042503 0.9538214
#> [10,] 1.29723498 2.00277317 0.7128106

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]        [,2]      [,3]
#>  [1,] 3.0946343 2.151024308 0.4511987
#>  [2,] 0.1124593 0.004720902 0.7896864
#>  [3,] 0.1614089 0.182892634 0.1828926
#>  [4,] 4.3434254 2.242129333 1.0103375
#>  [5,] 4.5113399 7.067920719 7.0075909
#>  [6,] 0.1235802 0.290478768 0.2904788
#>  [7,] 1.4995728 1.526735828 1.5267358
#>  [8,] 4.7727326 0.428736134 0.8281848
#>  [9,] 3.8189171 2.550940007 1.0641789
#> [10,] 0.8235179 0.306280515 0.3870058
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09769278 0.09769278 0.09769278
#>  [2,] 0.14735060 0.14735060 0.14735060
#>  [3,] 0.06652061 0.06652061 0.06652061
#>  [4,] 0.52763762 0.52763762 0.52763762
#>  [5,] 1.32606378 1.32606378 1.32606378
#>  [6,] 0.45143193 0.45143193 0.45143193
#>  [7,] 1.06554559 1.06554559 1.06554559
#>  [8,] 0.29848389 0.29848389 0.29848389
#>  [9,] 0.61838450 0.61838450 0.61838450
#> [10,] 2.42302575 2.42302575 2.42302575
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 2.2316993 1.3246724 0.38641922
#>  [2,] 0.7077239 0.7006156 1.38257311
#>  [3,] 0.6078413 0.4556398 1.48393751
#>  [4,] 0.0632202 0.1443864 0.05481531
#>  [5,] 0.3871925 0.4685145 0.54476806
#>  [6,] 1.0073144 0.6587258 1.67869005
#>  [7,] 1.4088364 0.6605092 1.01486646
#>  [8,] 0.7866781 0.2881921 0.06937590
#>  [9,] 0.2589513 0.0569693 0.50307918
#> [10,] 0.5236039 3.2053015 0.39152257

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 2.85245171 5.9615073 0.01487636
#>  [2,] 0.06705883 6.6284630 0.03075425
#>  [3,] 0.42117078 2.7908157 0.12594660
#>  [4,] 1.97623452 2.9046492 1.93562582
#>  [5,] 1.00096150 1.2759796 1.27597964
#>  [6,] 0.67650332 0.8447073 0.90308879
#>  [7,] 1.00428570 1.9445930 2.96297728
#>  [8,] 5.67511651 4.2401957 9.56224243
#>  [9,] 0.72392342 0.9041515 1.22572003
#> [10,] 1.95038236 1.1556236 1.95038236
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3535809 1.3535809 1.3535809
#>  [2,] 1.2156435 1.2156435 1.2156435
#>  [3,] 1.3743318 1.3743318 1.3743318
#>  [4,] 0.1837457 0.1837457 0.1837457
#>  [5,] 0.5012098 0.5012098 0.5012098
#>  [6,] 1.7017602 1.7017602 1.7017602
#>  [7,] 0.7191187 0.7191187 0.7191187
#>  [8,] 0.1733889 0.1733889 0.1733889
#>  [9,] 0.0188770 0.0188770 0.0188770
#> [10,] 1.0232438 1.0232438 1.0232438
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]      [,2]        [,3]
#>  [1,] 0.33845306 0.2883139 1.413391886
#>  [2,] 0.38826948 2.6618487 1.024707969
#>  [3,] 0.03381794 1.6948712 0.250536554
#>  [4,] 0.14528577 0.9493135 1.274531615
#>  [5,] 1.34547792 0.5316884 2.239321001
#>  [6,] 0.76596336 0.2375554 1.654812053
#>  [7,] 0.35932202 0.2262418 2.248101240
#>  [8,] 1.50658121 0.2642157 0.002029912
#>  [9,] 1.63985395 1.0477428 1.380314912
#> [10,] 0.22371648 1.0461123 1.035215187

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2415149 0.8380647 1.3970931
#>  [2,] 2.6815575 0.4385459 1.2634880
#>  [3,] 1.5396967 1.3641468 0.5311008
#>  [4,] 0.5128424 0.5128424 0.5128424
#>  [5,] 0.1921213 2.0489479 0.3822741
#>  [6,] 0.3913151 0.3913151 0.2642290
#>  [7,] 8.8596200 0.6450250 3.5283411
#>  [8,] 0.7693254 1.0520011 0.4459968
#>  [9,] 0.8883895 1.8045030 1.0020017
#> [10,] 0.5572188 3.4003406 1.2155034
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.60229219 1.60229219 1.60229219
#>  [2,] 0.01750405 0.01750405 0.01750405
#>  [3,] 2.13942247 2.13942247 2.13942247
#>  [4,] 4.28114383 4.28114383 4.28114383
#>  [5,] 4.18537489 4.18537489 4.18537489
#>  [6,] 0.93412888 0.93412888 0.93412888
#>  [7,] 2.06363748 2.06363748 2.06363748
#>  [8,] 0.05101009 0.05101009 0.05101009
#>  [9,] 0.12260154 0.12260154 0.12260154
#> [10,] 0.06979371 0.06979371 0.06979371
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 3.8799848 4.5274974 0.07913731
#>  [2,] 1.3652451 0.9165304 0.33278946
#>  [3,] 0.2437675 0.6765321 0.37793814
#>  [4,] 0.3692427 0.1604913 0.04657202
#>  [5,] 0.2878478 0.3222696 1.60774386
#>  [6,] 0.1428277 0.2322180 0.48598554
#>  [7,] 1.6203728 1.1551634 0.75365807
#>  [8,] 1.1665909 0.8139037 2.24003232
#>  [9,] 0.3926033 1.1032037 0.38960624
#> [10,] 0.7434654 0.3148305 0.76011851

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 4.65247004 0.46464703 0.8797730
#>  [2,] 0.58917572 0.58917572 0.5891757
#>  [3,] 0.21214629 0.07056846 0.5398551
#>  [4,] 0.37520462 1.76451684 1.7784255
#>  [5,] 0.66448314 0.26801484 0.7645120
#>  [6,] 0.05916599 0.22954146 1.2731541
#>  [7,] 0.41758225 0.28510247 3.4158693
#>  [8,] 1.91012294 0.82493858 1.6913781
#>  [9,] 1.67582893 0.67950524 3.1573068
#> [10,] 3.55914540 1.97277464 1.0080162
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7198395 0.7198395 0.7198395
#>  [2,] 0.8283675 0.8283675 0.8283675
#>  [3,] 1.7972816 1.7972816 1.7972816
#>  [4,] 0.1171490 0.1171490 0.1171490
#>  [5,] 0.6168948 0.6168948 0.6168948
#>  [6,] 2.1083604 2.1083604 2.1083604
#>  [7,] 0.5016181 0.5016181 0.5016181
#>  [8,] 1.0599512 1.0599512 1.0599512
#>  [9,] 0.7334609 0.7334609 0.7334609
#> [10,] 0.5510449 0.5510449 0.5510449
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.02441064 0.2611815 0.9934070
#>  [2,] 0.94019527 0.6326062 0.6366638
#>  [3,] 0.82152648 0.5798298 2.3084865
#>  [4,] 0.13732175 0.8294448 1.2878341
#>  [5,] 0.02686110 0.4337343 2.0919478
#>  [6,] 0.52658480 1.5946428 0.6289367
#>  [7,] 3.60998063 2.1755008 0.9839016
#>  [8,] 1.02385177 0.3320619 0.2520072
#>  [9,] 0.68762032 0.8183325 1.0715326
#> [10,] 1.78972464 1.8035619 0.3719364

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1290093 0.1290093 0.1290093
#>  [2,] 0.9840093 1.1993913 2.0251560
#>  [3,] 0.3629227 0.4104974 0.4239124
#>  [4,] 0.6776874 0.1041686 1.5537105
#>  [5,] 0.3138959 0.1434016 0.4491913
#>  [6,] 0.3772609 0.3772609 0.3772609
#>  [7,] 1.0153038 0.1148988 0.2630921
#>  [8,] 0.0984177 0.0984177 0.1038433
#>  [9,] 0.4960575 2.5960992 2.8422305
#> [10,] 0.6952040 1.3486394 0.8147379
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08169722 0.08169722 0.08169722
#>  [2,] 1.24866755 2.20989693 1.07186427
#>  [3,] 1.02562094 0.20241190 0.02167168
#>  [4,] 0.01378372 0.33213495 0.56232224
#>  [5,] 0.44866175 0.27393323 0.27507479
#>  [6,] 0.36529510 1.02127121 0.99785908
#>  [7,] 0.46226912 0.46226912 1.17317776
#>  [8,] 0.06744114 0.06744114 0.64051047
#>  [9,] 0.12007388 1.74489226 1.74489226
#> [10,] 0.16588503 0.53839113 0.01502760
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56122438 0.08319941 0.08319941
#>  [2,] 1.24286940 0.31982719 1.28258284
#>  [3,] 0.37057108 0.62493341 1.70523036
#>  [4,] 0.63416944 1.06364941 0.63485273
#>  [5,] 0.29256332 0.90992615 0.15537019
#>  [6,] 0.08098002 0.29786690 0.29786690
#>  [7,] 0.06162059 1.98907443 0.64794974
#>  [8,] 0.18287304 0.63882627 0.36965971
#>  [9,] 0.23852927 0.21102429 1.03699278
#> [10,] 0.73819483 0.11876766 1.56166088
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.81623823 0.8173103 0.81731028
#>  [2,] 0.86233294 0.6732523 1.08707405
#>  [3,] 2.21360211 1.0357773 0.64702405
#>  [4,] 0.29270831 0.6195707 0.58611653
#>  [5,] 0.17464257 0.1843455 0.17464257
#>  [6,] 0.55373836 0.0833659 1.17583257
#>  [7,] 0.37861214 0.9512186 0.23829632
#>  [8,] 0.09641278 0.5938083 0.70376855
#>  [9,] 0.01104785 1.0626989 0.01104785
#> [10,] 0.22362377 0.1744632 1.83636662
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34157133 0.10016756 0.09264861
#>  [2,] 0.23201670 0.23201670 0.23201670
#>  [3,] 0.87771507 1.84633278 0.90112476
#>  [4,] 0.71633890 0.55907612 0.71633890
#>  [5,] 0.22139794 0.76725982 0.22139794
#>  [6,] 0.11402205 0.71179934 0.71179934
#>  [7,] 0.01203128 0.26966051 1.81377253
#>  [8,] 1.96795203 0.28017611 1.16811204
#>  [9,] 0.34674402 0.06009245 0.67056147
#> [10,] 0.11531311 1.52766482 0.31338525

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2646165 0.8028737 0.4212599
#>  [2,] 1.2385906 4.2136611 1.1781516
#>  [3,] 0.1945427 3.7297479 2.0637957
#>  [4,] 0.3203371 0.3203371 0.3203371
#>  [5,] 0.4424193 0.4424193 0.4424193
#>  [6,] 0.4193171 0.4193171 0.4193171
#>  [7,] 1.8234968 1.8234968 1.8234968
#>  [8,] 0.1141714 0.1141714 0.1141714
#>  [9,] 2.2039282 0.3558110 0.2411948
#> [10,] 1.9736514 4.2806269 3.3611890
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.73292954 2.01898288 4.59065754
#>  [2,] 0.09605711 0.09605711 0.09605711
#>  [3,] 3.54166566 1.67207684 0.17091735
#>  [4,] 0.04086147 0.45258163 1.33658897
#>  [5,] 3.07596401 1.31578764 3.07596401
#>  [6,] 0.91858804 0.91858804 0.65681742
#>  [7,] 0.17604691 2.92371123 0.41175999
#>  [8,] 1.28835289 0.03839447 3.31755079
#>  [9,] 0.30787434 0.07602021 2.02218095
#> [10,] 5.22712867 0.52386574 0.72388452
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.9652714 0.00370196 1.1960835
#>  [2,] 2.1587332 0.96904397 1.0207437
#>  [3,] 0.7079483 5.16377871 4.7850954
#>  [4,] 0.5146495 0.51464952 0.5146495
#>  [5,] 4.6024294 4.31074808 0.5463381
#>  [6,] 0.1319518 4.66527982 1.3430389
#>  [7,] 0.5668254 0.56682536 0.5668254
#>  [8,] 0.1336137 1.30756154 0.7018605
#>  [9,] 1.9290536 0.43507462 3.2221751
#> [10,] 1.7933621 0.75616094 1.1489472
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]      [,2]        [,3]
#>  [1,] 2.27129621 2.2712962 0.118233316
#>  [2,] 1.86659417 0.2954246 0.760801811
#>  [3,] 0.95028988 0.9647283 0.202357015
#>  [4,] 0.65812414 0.3617035 0.658124145
#>  [5,] 2.72005384 0.4078272 0.005372346
#>  [6,] 0.01992557 2.0693150 0.720001658
#>  [7,] 0.12495749 1.8135585 1.640107354
#>  [8,] 1.00676787 0.3890623 0.939034670
#>  [9,] 2.53997422 1.0703266 0.158093283
#> [10,] 0.42449409 0.5256270 0.565185393
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4702259 0.4702259 0.4702259
#>  [2,] 1.4306797 0.2111784 3.7132225
#>  [3,] 0.1094266 1.3543393 0.7302321
#>  [4,] 0.7073869 0.5893672 2.7923292
#>  [5,] 0.5379032 0.4676020 1.5921512
#>  [6,] 1.6946128 1.8203555 1.3540194
#>  [7,] 0.4375105 0.2096792 0.3472753
#>  [8,] 0.1775563 3.1738273 0.1474925
#>  [9,] 2.0627273 1.7032284 0.1512586
#> [10,] 0.7670360 0.5900380 4.8189034

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.91385932 1.51689402 1.14702698
#>  [2,] 1.16792804 1.16792804 1.16792804
#>  [3,] 0.08451309 0.08451309 0.08451309
#>  [4,] 0.21953087 0.21953087 0.21953087
#>  [5,] 1.22157323 1.22157323 0.66918775
#>  [6,] 0.32859426 0.84346599 0.32859426
#>  [7,] 0.13825836 0.93449376 0.13825836
#>  [8,] 0.14110710 0.25837544 1.21834728
#>  [9,] 0.03034994 0.04170635 0.04170635
#> [10,] 0.22234943 0.07219484 0.07219484
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.71854909 0.01037376 0.93096511
#>  [2,] 0.39781824 0.93250898 0.52206764
#>  [3,] 0.02266552 0.02266552 0.02266552
#>  [4,] 0.16215608 0.87613178 0.16215608
#>  [5,] 1.09785011 1.05334670 1.09785011
#>  [6,] 0.65263327 0.65263327 0.65263327
#>  [7,] 1.09362110 1.09362110 0.38129008
#>  [8,] 2.01362081 1.86748188 1.92503074
#>  [9,] 0.33130060 0.04479560 0.03789566
#> [10,] 0.60551488 0.60551488 0.60551488
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.778053009 0.77805301 0.62237435
#>  [2,] 0.091346657 0.09134666 0.09134666
#>  [3,] 1.390443259 0.86373113 1.07265723
#>  [4,] 0.104615075 0.10461508 0.10461508
#>  [5,] 0.867077913 0.86707791 0.86707791
#>  [6,] 0.959918364 0.95991836 0.95991836
#>  [7,] 0.005565635 1.82463605 0.55513049
#>  [8,] 1.478306986 0.15160881 0.37236356
#>  [9,] 0.541467501 0.01615954 0.54146750
#> [10,] 0.201007589 0.20100759 0.17654864
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.0735223 0.18904040 0.18904040
#>  [2,] 0.5903578 0.37465930 0.37465930
#>  [3,] 0.3874635 0.45465483 0.38746349
#>  [4,] 0.2813260 0.28132605 0.28132605
#>  [5,] 1.4914472 0.06773978 1.49144719
#>  [6,] 0.4417252 0.44172516 0.03846859
#>  [7,] 0.4356764 0.43567644 0.43567644
#>  [8,] 0.4335665 0.23882742 0.23882742
#>  [9,] 1.2276306 1.22763061 0.69913660
#> [10,] 0.1633272 0.16332715 0.16332715
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.92795252 0.92795252 0.488498809
#>  [2,] 0.58697114 0.58697114 0.586971141
#>  [3,] 0.07287409 0.07287409 0.072874088
#>  [4,] 1.24708849 0.01923661 0.801244957
#>  [5,] 0.27358731 0.02998281 0.029982809
#>  [6,] 0.01412552 0.70860930 0.004336575
#>  [7,] 0.13849361 0.13849361 0.087110766
#>  [8,] 0.76566497 0.94248327 1.127671551
#>  [9,] 0.09599950 0.09030311 0.095999505
#> [10,] 0.39199793 0.39199793 0.391997925

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12482483 0.60994320 0.60994320
#>  [2,] 0.26200148 0.26200148 0.26200148
#>  [3,] 0.29267465 0.09865515 0.29267465
#>  [4,] 0.49890307 0.49590478 1.18165983
#>  [5,] 0.08588699 0.08588699 0.01709060
#>  [6,] 0.03942242 0.03942242 0.03942242
#>  [7,] 0.03541137 0.03541137 0.03541137
#>  [8,] 0.20871685 0.20871685 0.20871685
#>  [9,] 0.14063032 0.49072086 0.14063032
#> [10,] 0.50598617 0.50598617 0.50598617
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.1888465 1.1156187 1.18884648
#>  [2,] 0.1190308 0.1190308 0.11903082
#>  [3,] 0.1287550 0.1287550 0.12875504
#>  [4,] 0.1192430 0.3802062 0.11924301
#>  [5,] 1.1276497 1.2102790 1.21027904
#>  [6,] 0.3757946 0.1971527 0.01868955
#>  [7,] 0.4792309 0.4792309 0.11637077
#>  [8,] 0.7453607 0.1568359 0.14851353
#>  [9,] 1.3739824 0.1329812 0.88864607
#> [10,] 0.2592039 0.2592039 0.64034960
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15826129 0.94693125 0.94693125
#>  [2,] 0.63487826 0.27871092 0.91542523
#>  [3,] 0.60598017 0.84292201 0.22414132
#>  [4,] 0.34776149 0.34776149 0.34776149
#>  [5,] 0.27634541 0.27634541 0.27634541
#>  [6,] 0.31443842 0.08348688 0.31443842
#>  [7,] 0.47603053 0.40863194 0.06615780
#>  [8,] 0.17742546 0.02460506 0.02460506
#>  [9,] 0.27685285 0.01613443 0.01613443
#> [10,] 0.04951015 0.07035604 0.15965900
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11088903 0.11088903 0.11088903
#>  [2,] 0.26469080 0.31945116 0.10231289
#>  [3,] 0.28291754 1.12826887 1.12826887
#>  [4,] 0.07119451 0.12671684 0.12313323
#>  [5,] 0.40791243 0.08626181 0.43239893
#>  [6,] 0.11067678 1.19822483 0.15203131
#>  [7,] 0.10631554 0.37483429 0.72367043
#>  [8,] 0.68905228 0.18107903 0.50257896
#>  [9,] 0.10866604 0.10866604 0.01245902
#> [10,] 0.02024315 0.02024315 0.02024315

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>              [,1]         [,2]         [,3]
#>  [1,] 0.414718146 0.5735730374 0.1314179192
#>  [2,] 0.963263846 0.4453264879 0.4453264879
#>  [3,] 0.224385219 0.0005088598 0.0005088598
#>  [4,] 0.232486287 0.2324862870 0.2324862870
#>  [5,] 0.006522083 0.8982160700 0.0466478540
#>  [6,] 0.729482128 0.7294821278 0.9727236454
#>  [7,] 0.277328295 0.2773282947 0.1769023795
#>  [8,] 0.245404195 0.7063309361 1.6532526512
#>  [9,] 0.060606309 0.3129120240 0.3658047291
#> [10,] 1.191634127 0.2266003927 0.2501870937
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06293327 0.54102213 0.19238646
#>  [2,] 0.20619281 0.20510148 0.20619281
#>  [3,] 0.69986871 0.69986871 0.69986871
#>  [4,] 0.28993911 0.89689724 0.05992806
#>  [5,] 0.02991882 0.16470414 0.16470414
#>  [6,] 0.55530604 0.10421721 0.10421721
#>  [7,] 0.28490337 0.28490337 0.28490337
#>  [8,] 0.47045419 0.69193502 0.38901866
#>  [9,] 0.04365089 0.08575455 0.09343196
#> [10,] 0.78422956 0.44151878 0.16531914
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>            [,1]        [,2]        [,3]
#>  [1,] 0.7289443 0.364369320 0.295859391
#>  [2,] 0.4607430 0.033190441 0.411401020
#>  [3,] 0.6284136 0.257040898 0.628413604
#>  [4,] 0.1416925 0.171734935 0.150167186
#>  [5,] 1.1159029 0.147114042 0.147114042
#>  [6,] 1.1906325 0.203165983 0.489444843
#>  [7,] 0.4693050 0.004456675 0.004456675
#>  [8,] 0.1435802 0.095729516 0.152340204
#>  [9,] 0.2633616 0.931941098 0.931941098
#> [10,] 1.7210409 0.465016055 0.070672838
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.2075947 1.72207614 0.18952435
#>  [2,] 0.2312972 0.19487925 0.19487925
#>  [3,] 0.4513560 0.87507352 0.20121644
#>  [4,] 0.2984166 0.29841659 0.29841659
#>  [5,] 0.2988533 0.38651809 0.14097492
#>  [6,] 0.3414812 0.02694526 0.31155510
#>  [7,] 0.5945899 0.59458989 1.15884557
#>  [8,] 0.5106004 0.16795378 0.25848452
#>  [9,] 0.3551737 0.35517368 0.02347336
#> [10,] 0.7567248 0.33799028 0.17706144

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.27026540 0.25385378 0.1588377
#>  [2,] 0.19075832 0.19075832 0.1907583
#>  [3,] 0.38927185 0.40490075 0.3892719
#>  [4,] 0.33755099 0.33930754 0.3393075
#>  [5,] 0.10233082 0.10233082 0.1023308
#>  [6,] 0.26505890 1.18629891 0.7173327
#>  [7,] 0.07822588 0.07822588 0.5990458
#>  [8,] 0.10572170 0.10572170 0.1616733
#>  [9,] 0.43302375 0.09313376 0.1583385
#> [10,] 0.11911473 0.24291278 0.2429128
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.21378354 1.1112166 0.76656650
#>  [2,] 0.13421409 0.9407272 0.30097357
#>  [3,] 0.20360937 0.2036094 0.20360937
#>  [4,] 0.93317630 0.2331870 0.93317630
#>  [5,] 0.09190412 0.3004305 0.09190412
#>  [6,] 0.18923899 0.3859030 0.38590298
#>  [7,] 0.48627101 0.4862710 0.37102292
#>  [8,] 0.78160727 0.7816073 0.01439485
#>  [9,] 0.72877153 0.1271398 0.89788286
#> [10,] 0.18379635 0.1837964 0.18379635
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1597580 0.5041695 0.2143300
#>  [2,] 0.1817840 0.1817840 0.1817840
#>  [3,] 0.2898676 0.3094860 0.3094860
#>  [4,] 0.1536863 0.1536863 1.6374207
#>  [5,] 0.1212746 0.2205054 0.5320312
#>  [6,] 0.5282591 0.5282591 1.0665644
#>  [7,] 0.5511841 0.9448607 0.1253771
#>  [8,] 0.1119215 0.2423467 0.2423467
#>  [9,] 0.6260432 0.1172253 0.2056855
#> [10,] 0.1935302 0.5074447 0.2143436
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 0.1499967 0.283867540 0.28386754
#>  [2,] 1.0020627 0.006728227 1.04903887
#>  [3,] 0.7925607 0.403890780 0.97866420
#>  [4,] 0.1030193 0.050333196 0.05033320
#>  [5,] 0.2293041 0.229304089 0.99034578
#>  [6,] 0.7196549 0.007750693 1.14421771
#>  [7,] 0.4543194 0.052537582 0.05253758
#>  [8,] 0.1551611 0.155161082 0.15516108
#>  [9,] 0.1881755 0.188175457 0.18817546
#> [10,] 0.2093521 0.234375620 0.23437562