Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8808379 0.6798315 2.96510313
#>  [2,] 5.1864943 0.5658230 0.74377280
#>  [3,] 1.4633846 0.8080718 1.46338461
#>  [4,] 0.5133795 0.5133795 0.51337950
#>  [5,] 2.0464852 4.0307445 0.63914462
#>  [6,] 0.3048157 2.5723129 1.21290924
#>  [7,] 1.6204940 1.7659800 1.76597999
#>  [8,] 0.3311425 2.8498172 0.18622547
#>  [9,] 1.5436456 1.5436456 0.03685121
#> [10,] 2.0129310 2.0129310 2.01293096
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21371976 0.21371976 0.21371976
#>  [2,] 0.18928452 0.18928452 0.18928452
#>  [3,] 0.80353204 0.80353204 0.80353204
#>  [4,] 0.99208825 0.99208825 0.99208825
#>  [5,] 0.95870776 0.95870776 0.95870776
#>  [6,] 1.56381393 1.56381393 1.56381393
#>  [7,] 0.70728257 0.70728257 0.70728257
#>  [8,] 0.04636188 0.04636188 0.04636188
#>  [9,] 0.42414301 0.42414301 0.42414301
#> [10,] 0.60955485 0.60955485 0.60955485
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17529442 0.05181549 0.80034837
#>  [2,] 0.02022363 5.15129148 0.42276863
#>  [3,] 0.25869322 2.17387787 0.06041065
#>  [4,] 1.27198030 0.44400032 1.60899688
#>  [5,] 0.01529083 4.22404651 0.49176144
#>  [6,] 1.99984931 0.65595199 0.39467390
#>  [7,] 2.95477921 3.18578450 1.82686908
#>  [8,] 1.10282343 0.22772918 0.11642614
#>  [9,] 1.67663521 0.24826603 1.37104034
#> [10,] 1.04623717 2.34602492 0.43198431

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,]  0.1262786 0.1262786 0.1262786
#>  [2,]  4.8335412 1.2255042 0.4961422
#>  [3,]  3.1654084 0.3617878 0.3136416
#>  [4,]  0.2522335 0.2522335 0.2522335
#>  [5,]  0.1768718 0.1768718 0.1768718
#>  [6,] 10.5157801 0.1104382 1.0357391
#>  [7,]  4.8769426 0.6248069 0.3775621
#>  [8,]  0.2260654 0.9295262 1.2602722
#>  [9,]  4.1900481 0.4623095 1.6327583
#> [10,]  0.2958036 2.1223618 0.8526738
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0543108 2.0543108 2.0543108
#>  [2,] 1.0449293 1.0449293 1.0449293
#>  [3,] 0.6679115 0.6679115 0.6679115
#>  [4,] 0.1778765 0.1778765 0.1778765
#>  [5,] 1.9905592 1.9905592 1.9905592
#>  [6,] 0.1592025 0.1592025 0.1592025
#>  [7,] 0.4651083 0.4651083 0.4651083
#>  [8,] 0.3367281 0.3367281 0.3367281
#>  [9,] 3.5324615 3.5324615 3.5324615
#> [10,] 0.5671389 0.5671389 0.5671389
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41129760 0.01738931 3.17184468
#>  [2,] 0.59020265 1.30670763 1.20628023
#>  [3,] 0.32961096 0.21052182 2.76158344
#>  [4,] 1.35258866 0.08904706 1.25271127
#>  [5,] 0.04607793 0.85441434 2.15642923
#>  [6,] 0.90880205 0.36479888 1.86602057
#>  [7,] 0.77957018 1.67188179 0.65174581
#>  [8,] 1.70560206 1.14463158 1.40361502
#>  [9,] 0.16186364 0.46132847 0.48181977
#> [10,] 0.03180200 1.15395291 0.05270253

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>              [,1]        [,2]        [,3]
#>  [1,] 1.889834707 0.457991078 1.889834707
#>  [2,] 1.382884791 2.912822278 1.960653076
#>  [3,] 0.627384249 0.855516738 1.963548183
#>  [4,] 0.520938798 0.134710958 1.695862627
#>  [5,] 0.705743037 0.770403060 0.506707165
#>  [6,] 0.004337629 0.004337629 0.004337629
#>  [7,] 5.142523389 1.002329364 8.513730980
#>  [8,] 0.155350887 0.061390176 1.572896775
#>  [9,] 0.397607906 0.372424956 0.173869765
#> [10,] 1.381855822 0.614835378 2.135015869
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.96320170 0.96320170 0.96320170
#>  [2,] 0.38892757 0.38892757 0.38892757
#>  [3,] 2.83083019 2.83083019 2.83083019
#>  [4,] 1.35484069 1.35484069 1.35484069
#>  [5,] 0.05176067 0.05176067 0.05176067
#>  [6,] 0.38066872 0.38066872 0.38066872
#>  [7,] 0.81192075 0.81192075 0.81192075
#>  [8,] 0.02506405 0.02506405 0.02506405
#>  [9,] 1.34013909 1.34013909 1.34013909
#> [10,] 0.41913967 0.41913967 0.41913967
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.01365027 1.1445519 0.20584599
#>  [2,] 1.21063580 0.5711351 0.14325199
#>  [3,] 4.60428578 0.7134082 0.89546561
#>  [4,] 0.47449393 1.2307055 0.27155614
#>  [5,] 0.05245863 0.4383928 0.08128495
#>  [6,] 0.43934629 1.5821385 0.80983716
#>  [7,] 0.53782971 0.4501785 0.43464929
#>  [8,] 1.16201051 1.7255416 2.48328039
#>  [9,] 1.73961978 1.8657484 0.67278301
#> [10,] 0.16630583 0.5324279 2.62362938

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.09022612 2.9371301 0.3464499
#>  [2,] 3.80495451 2.7823532 0.8370931
#>  [3,] 0.20772742 0.5566553 1.9458313
#>  [4,] 1.58555992 1.5855599 0.9046711
#>  [5,] 0.55455144 0.3552920 0.4633187
#>  [6,] 0.15881701 0.1588170 0.1588170
#>  [7,] 0.76774735 0.7677474 0.6697970
#>  [8,] 7.25520770 3.5882410 2.9478919
#>  [9,] 0.04272174 0.3104633 1.5248946
#> [10,] 2.06568906 0.9791008 1.0525245
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.49262873 0.49262873 0.49262873
#>  [2,] 1.12200100 1.12200100 1.12200100
#>  [3,] 0.36570533 0.36570533 0.36570533
#>  [4,] 1.05876188 1.05876188 1.05876188
#>  [5,] 0.33780852 0.33780852 0.33780852
#>  [6,] 1.57496971 1.57496971 1.57496971
#>  [7,] 0.02854741 0.02854741 0.02854741
#>  [8,] 1.67085980 1.67085980 1.67085980
#>  [9,] 0.16010975 0.16010975 0.16010975
#> [10,] 1.52256529 1.52256529 1.52256529
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09876632 1.17925248 0.07207066
#>  [2,] 0.99391252 0.20485211 0.53382194
#>  [3,] 0.14629311 0.19173699 0.42645243
#>  [4,] 1.10478438 0.21734432 0.26947996
#>  [5,] 0.11537501 0.49458154 0.26745761
#>  [6,] 1.46783724 0.00893565 1.96593324
#>  [7,] 2.06523958 3.23502786 2.30914003
#>  [8,] 1.03073354 1.00091330 0.78441533
#>  [9,] 0.09311004 1.00795335 0.10474979
#> [10,] 2.53814971 0.41077884 0.88189109

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.308316658 0.3083167 0.3083167
#>  [2,] 1.133167803 0.8851606 0.9965718
#>  [3,] 1.388819938 0.1943893 1.4503761
#>  [4,] 0.757747643 2.1992053 1.1838549
#>  [5,] 1.096622024 0.8716364 0.7453277
#>  [6,] 1.202117212 1.3393270 4.0781087
#>  [7,] 0.004249763 0.4165302 3.6987903
#>  [8,] 4.024343127 2.7081752 1.9810935
#>  [9,] 1.076089469 4.7952613 1.7520252
#> [10,] 2.955764311 0.6978507 2.3640290
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.74421196 0.74421196 0.74421196
#>  [2,] 0.30112041 0.30112041 0.30112041
#>  [3,] 1.83615567 1.83615567 1.83615567
#>  [4,] 1.32857425 1.32857425 1.32857425
#>  [5,] 0.07926102 0.07926102 0.07926102
#>  [6,] 0.34694547 0.34694547 0.34694547
#>  [7,] 0.44035627 0.44035627 0.44035627
#>  [8,] 1.29620596 1.29620596 1.29620596
#>  [9,] 0.32086048 0.32086048 0.32086048
#> [10,] 3.10365251 3.10365251 3.10365251
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 5.23029380 1.20736368 0.92993973
#>  [2,] 0.47021739 1.64312742 0.01801413
#>  [3,] 2.93020171 0.21321578 2.49977595
#>  [4,] 2.54976546 0.16690136 2.66077969
#>  [5,] 2.36603335 1.51714297 1.21600176
#>  [6,] 0.04974313 0.49200423 0.31550769
#>  [7,] 0.19635333 0.36826173 0.28215593
#>  [8,] 0.02293908 0.04693744 0.19103539
#>  [9,] 0.20334325 0.22762153 0.36435797
#> [10,] 2.21077469 0.98557752 0.28395749

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 4.33345589 3.4742328 0.2869648
#>  [2,] 1.00113718 4.5644371 1.7800779
#>  [3,] 0.09047205 2.5686027 1.1411100
#>  [4,] 2.97410128 2.5320924 2.9741013
#>  [5,] 2.57731571 0.1848100 4.9213322
#>  [6,] 2.12542521 1.6697408 2.7728455
#>  [7,] 0.38759367 0.2598084 2.8840425
#>  [8,] 2.79874305 2.3801130 2.7987430
#>  [9,] 3.93932666 1.0164784 0.7943467
#> [10,] 0.54988391 0.5498839 0.5498839
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5124184 0.5124184 0.5124184
#>  [2,] 1.9352706 1.9352706 1.9352706
#>  [3,] 0.4734460 0.4734460 0.4734460
#>  [4,] 1.0747194 1.0747194 1.0747194
#>  [5,] 3.6936908 3.6936908 3.6936908
#>  [6,] 2.5321563 2.5321563 2.5321563
#>  [7,] 0.1629573 0.1629573 0.1629573
#>  [8,] 0.8698979 0.8698979 0.8698979
#>  [9,] 0.2508632 0.2508632 0.2508632
#> [10,] 0.9606173 0.9606173 0.9606173
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.40979389 0.7730692 1.70500402
#>  [2,] 0.10665395 1.6837257 1.10137736
#>  [3,] 0.64712947 0.4700971 0.28696540
#>  [4,] 0.29154491 2.3999365 0.52449618
#>  [5,] 1.29673690 0.3417954 0.26794465
#>  [6,] 2.27488074 1.6141182 3.77950905
#>  [7,] 0.14882173 0.5198612 0.79601974
#>  [8,] 2.74715483 0.9679908 1.53003731
#>  [9,] 0.07977586 0.4542312 0.39088997
#> [10,] 0.07982632 2.3278324 0.07765818

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>              [,1]      [,2]       [,3]
#>  [1,] 0.775922381 0.7759224 0.77592238
#>  [2,] 3.661899950 0.1796252 0.77489754
#>  [3,] 0.005343117 0.5388334 0.25327425
#>  [4,] 0.297000807 0.2970008 0.29700081
#>  [5,] 2.806434043 3.3249063 1.67098545
#>  [6,] 0.101453955 1.4652354 0.06837628
#>  [7,] 0.357662970 0.3576630 0.35766297
#>  [8,] 0.276001212 0.6153765 3.46771331
#>  [9,] 3.030277388 1.2330170 0.40471189
#> [10,] 0.085999712 1.0560249 1.05602494
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36299648 0.36299648 0.36299648
#>  [2,] 1.02147932 1.02147932 1.02147932
#>  [3,] 0.49332763 0.49332763 0.49332763
#>  [4,] 0.75762402 0.75762402 0.75762402
#>  [5,] 0.10530382 0.10530382 0.10530382
#>  [6,] 0.75251179 0.75251179 0.75251179
#>  [7,] 1.34589545 1.34589545 1.34589545
#>  [8,] 0.37908736 0.37908736 0.37908736
#>  [9,] 1.41066259 1.41066259 1.41066259
#> [10,] 0.08549341 0.08549341 0.08549341
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 2.69139231 0.07373169 0.9190609
#>  [2,] 0.42580630 4.62554784 0.2550922
#>  [3,] 0.06322691 1.09390218 1.0431391
#>  [4,] 0.58650062 1.06086376 0.3399598
#>  [5,] 0.09322318 0.56256946 0.9365158
#>  [6,] 3.66830954 1.02824873 0.2385290
#>  [7,] 1.14302482 5.35585372 2.2344020
#>  [8,] 0.70718285 2.90033904 0.6530477
#>  [9,] 2.09310327 0.91835968 0.3520242
#> [10,] 0.54007772 0.27113857 0.7116585

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.14283614 0.1428361 0.10500765
#>  [2,] 0.78493063 1.3139705 2.12142383
#>  [3,] 4.52889063 2.0779540 0.05413199
#>  [4,] 0.05369407 0.9012949 0.90129490
#>  [5,] 0.65302029 3.8387450 0.46041637
#>  [6,] 1.26611654 0.7563333 5.74521649
#>  [7,] 0.63964294 0.9149257 0.91492575
#>  [8,] 2.43254597 2.4325460 0.40605062
#>  [9,] 0.15889353 0.1588935 0.15889353
#> [10,] 1.35363033 0.1089486 0.12716370
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.7656194 1.7656194 1.7656194
#>  [2,] 0.5642448 0.5642448 0.5642448
#>  [3,] 0.4796806 0.4796806 0.4796806
#>  [4,] 1.0445270 1.0445270 1.0445270
#>  [5,] 0.7721420 0.7721420 0.7721420
#>  [6,] 0.2369055 0.2369055 0.2369055
#>  [7,] 2.3130227 2.3130227 2.3130227
#>  [8,] 0.3163999 0.3163999 0.3163999
#>  [9,] 0.7441080 0.7441080 0.7441080
#> [10,] 1.9332852 1.9332852 1.9332852
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.74152816 0.88053672 0.86588041
#>  [2,] 0.27530308 0.04252987 1.08511265
#>  [3,] 0.49789615 0.19761013 0.12267977
#>  [4,] 1.01764191 0.23491137 0.06275031
#>  [5,] 0.19713756 1.01086524 0.38423274
#>  [6,] 0.05997516 1.09323954 0.08652274
#>  [7,] 0.48899824 0.64004378 1.09533360
#>  [8,] 0.01246000 0.90345749 0.23921847
#>  [9,] 0.63222125 0.13342617 1.97399353
#> [10,] 0.59332933 0.60539846 0.90761410

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.27335051 3.70084104 0.38903298
#>  [2,] 1.08175246 1.08175246 1.08175246
#>  [3,] 0.73890599 0.73890599 0.29027443
#>  [4,] 2.93186527 0.45989161 0.16695299
#>  [5,] 1.33430205 2.49400550 0.95523347
#>  [6,] 0.77826191 0.41858725 0.34487745
#>  [7,] 0.38195738 0.38195738 0.38195738
#>  [8,] 0.02026983 0.02026983 0.02026983
#>  [9,] 4.13673054 1.25199190 0.86985268
#> [10,] 0.08466491 0.13758046 0.13758046
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15139546 0.15139546 0.15139546
#>  [2,] 0.01824237 0.01824237 0.01824237
#>  [3,] 0.83848984 0.83848984 0.83848984
#>  [4,] 1.19439640 1.19439640 1.19439640
#>  [5,] 0.80304007 0.80304007 0.80304007
#>  [6,] 0.68850770 0.68850770 0.68850770
#>  [7,] 0.77618278 0.77618278 0.77618278
#>  [8,] 0.92255221 0.92255221 0.92255221
#>  [9,] 1.50476422 1.50476422 1.50476422
#> [10,] 0.93426843 0.93426843 0.93426843
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]         [,2]       [,3]
#>  [1,] 0.2191307 0.0003657642 0.23702519
#>  [2,] 0.5805424 0.1175582393 0.05106003
#>  [3,] 0.7458176 2.2751520847 0.38019428
#>  [4,] 0.4964530 0.0089423504 1.54867775
#>  [5,] 2.9981574 0.9781095657 0.13422583
#>  [6,] 0.6490423 0.8853832909 0.19637270
#>  [7,] 0.9085245 1.8727533278 0.74468204
#>  [8,] 1.0699216 0.1845539909 1.48337882
#>  [9,] 0.4081772 0.2614102773 1.46619758
#> [10,] 0.6024914 1.7833460148 0.07486165

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2209797 2.3572088 2.3572088
#>  [2,] 0.2181837 0.4372869 0.4372869
#>  [3,] 0.4108560 0.6956921 0.4321552
#>  [4,] 0.8576078 0.3519111 1.5821067
#>  [5,] 0.3504981 0.7257890 4.5599540
#>  [6,] 0.4682282 0.4682282 0.4682282
#>  [7,] 1.6736957 0.5962695 2.6156896
#>  [8,] 2.5606355 2.5606355 2.5606355
#>  [9,] 1.4558700 0.6469146 0.2658445
#> [10,] 0.4996252 0.4996252 0.2784494
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27749496 0.27749496 0.27749496
#>  [2,] 0.05152759 0.05152759 0.05152759
#>  [3,] 2.07969471 2.07969471 2.07969471
#>  [4,] 2.16078862 2.16078862 2.16078862
#>  [5,] 0.31108590 0.31108590 0.31108590
#>  [6,] 0.82627740 0.82627740 0.82627740
#>  [7,] 0.20275203 0.20275203 0.20275203
#>  [8,] 0.18502860 0.18502860 0.18502860
#>  [9,] 0.17954091 0.17954091 0.17954091
#> [10,] 1.40091832 1.40091832 1.40091832
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 2.0427286 0.30102286 2.0389895
#>  [2,] 0.9424823 0.09181273 1.1421328
#>  [3,] 0.4118756 0.94882334 1.0990705
#>  [4,] 2.1070697 0.14730671 0.8996208
#>  [5,] 0.3242005 0.28765187 4.3414366
#>  [6,] 0.4495480 0.59041969 0.7220707
#>  [7,] 0.3945848 0.92039499 0.1099761
#>  [8,] 3.6754135 2.78588305 0.3469339
#>  [9,] 0.9012145 3.14204576 0.5996101
#> [10,] 1.8485676 2.32330614 4.3406049

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20845790 0.27535431 0.27535431
#>  [2,] 0.97878385 0.35474905 1.10277599
#>  [3,] 0.07959987 0.07959987 0.07959987
#>  [4,] 0.60418808 2.70892834 0.27375975
#>  [5,] 0.08134026 0.04241019 0.08134026
#>  [6,] 0.90094579 0.42137261 0.90094579
#>  [7,] 2.12239313 0.41337483 0.87928489
#>  [8,] 0.33913561 0.57910399 0.52895365
#>  [9,] 1.11397769 1.11397769 0.28262577
#> [10,] 1.30955884 0.52492670 0.34877549
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1817090 1.1347559 0.3019581
#>  [2,] 0.3271944 0.3271944 0.3271944
#>  [3,] 0.1239985 0.1239985 0.1239985
#>  [4,] 1.0440809 1.0440809 0.1538176
#>  [5,] 0.2047511 0.6836445 2.4915597
#>  [6,] 0.6666638 0.6284794 2.2356101
#>  [7,] 0.2881506 0.2881506 0.2881506
#>  [8,] 0.1178686 1.1456326 0.7432373
#>  [9,] 0.8131505 0.8131505 0.8131505
#> [10,] 0.4691389 1.0733965 0.2238727
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.07859268 2.23148273 1.3208888
#>  [2,] 0.25365529 1.81237005 0.5518459
#>  [3,] 0.33717062 0.07545700 0.1417392
#>  [4,] 1.46938201 2.16226166 1.0717545
#>  [5,] 0.07749442 0.07749442 1.0096892
#>  [6,] 0.18217860 0.12802770 0.5590366
#>  [7,] 0.75534724 0.37649151 0.7553472
#>  [8,] 0.11254967 0.11254967 0.1125497
#>  [9,] 1.90911606 0.15307994 1.4458493
#> [10,] 0.25966101 0.28985618 0.2898562
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.1944531 0.75841119 0.003209294
#>  [2,] 0.2489835 0.22263833 0.558868978
#>  [3,] 0.1357939 1.26285068 0.135793922
#>  [4,] 3.4414807 0.30479329 0.506167476
#>  [5,] 1.0551768 1.08578163 1.091243868
#>  [6,] 0.4062075 0.41634069 0.347738822
#>  [7,] 0.2016140 0.06787016 0.056468578
#>  [8,] 0.5354056 1.13595805 0.856268095
#>  [9,] 0.1398949 0.13989487 0.051413072
#> [10,] 0.9076994 0.13821494 0.719076522
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.8381576 0.47732663 0.838157555
#>  [2,] 0.6344893 0.17224881 0.578130232
#>  [3,] 1.4343750 0.32224079 0.663213407
#>  [4,] 0.7341891 0.73418912 0.340224395
#>  [5,] 0.4641166 0.54683029 0.165156151
#>  [6,] 0.6909431 0.03254818 0.032548177
#>  [7,] 0.7270701 0.50359553 0.002781682
#>  [8,] 0.3854833 0.75453265 0.194002826
#>  [9,] 0.5602559 0.46498939 0.560255861
#> [10,] 0.5370256 1.27879954 1.278799539

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8025464 1.2822131 1.0822587
#>  [2,] 0.9391949 0.5572753 2.1572245
#>  [3,] 1.9318246 0.8570769 4.2050262
#>  [4,] 0.2914025 0.2165818 0.1625830
#>  [5,] 1.1180200 1.1180200 0.3863334
#>  [6,] 2.3450139 0.1892559 1.1462819
#>  [7,] 0.7282192 0.7834981 0.9267715
#>  [8,] 0.3981620 0.2997987 0.3981620
#>  [9,] 0.1382368 2.9255248 2.9255248
#> [10,] 0.3743064 0.9375022 0.1624187
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.803201930 0.803201930 0.803201930
#>  [2,] 3.194739422 2.910382909 0.616124255
#>  [3,] 0.052695396 0.052695396 0.052695396
#>  [4,] 0.177689125 1.650107660 1.855521887
#>  [5,] 0.280741085 0.280741085 0.280741085
#>  [6,] 0.804843791 0.494337449 1.917204850
#>  [7,] 0.608121543 0.430518457 1.748863364
#>  [8,] 2.084549550 1.305605743 0.421994060
#>  [9,] 0.003920392 0.003920392 0.003920392
#> [10,] 2.991909559 5.424832257 2.246768262
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56123259 1.09497764 3.97866746
#>  [2,] 0.46065488 0.46065488 0.46065488
#>  [3,] 0.10944453 3.68975867 0.39997447
#>  [4,] 0.46568960 2.85190746 2.53294836
#>  [5,] 1.64442095 5.57099116 1.94355302
#>  [6,] 0.08262373 0.08262373 0.08262373
#>  [7,] 1.87453595 0.06373333 1.62876737
#>  [8,] 0.43734118 1.46364743 1.28793292
#>  [9,] 1.25698213 4.55076727 0.70259562
#> [10,] 0.02867790 0.02867790 0.02867790
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09824805 0.09824805 0.04259171
#>  [2,] 5.59259148 2.67742690 1.16668764
#>  [3,] 0.32148004 0.02265284 4.05340737
#>  [4,] 0.63503515 0.63503515 0.63503515
#>  [5,] 3.61847609 0.23814617 2.85772073
#>  [6,] 0.48408698 0.48408698 0.48408698
#>  [7,] 0.54310837 2.12905574 2.12905574
#>  [8,] 0.52449248 2.09899295 2.09899295
#>  [9,] 0.43567254 0.43567254 0.43567254
#> [10,] 0.78443911 2.68949882 1.85573026
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.34850406 1.3485041 1.34850406
#>  [2,] 0.90674082 0.1171097 0.08511536
#>  [3,] 0.20469344 0.2046934 0.20469344
#>  [4,] 0.31527738 1.6271241 0.74944220
#>  [5,] 1.43312265 1.2045175 1.43312265
#>  [6,] 0.10468371 0.1046837 0.10468371
#>  [7,] 1.17163873 1.1976450 1.19764503
#>  [8,] 0.01911578 3.7111172 5.16184170
#>  [9,] 0.42943718 0.4294372 0.42943718
#> [10,] 0.99955274 0.2482458 2.18274808

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]        [,2]      [,3]
#>  [1,] 0.14538741 0.145387412 0.1453874
#>  [2,] 0.21797342 0.217973416 0.2179734
#>  [3,] 0.42891184 0.574168840 0.6863543
#>  [4,] 0.03720121 0.007218653 0.2601355
#>  [5,] 1.15801384 1.158013838 1.1580138
#>  [6,] 0.21240849 0.212408492 0.1614498
#>  [7,] 0.17450279 0.342955276 1.2200849
#>  [8,] 0.57691850 0.576918502 0.5769185
#>  [9,] 0.18000716 0.099516215 0.1800072
#> [10,] 0.29427511 0.294275112 0.2942751
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>               [,1]       [,2]        [,3]
#>  [1,] 1.2292510520 0.11739424 0.667330633
#>  [2,] 0.0002733722 0.10075527 0.073017649
#>  [3,] 0.1676347645 0.16763476 0.167634765
#>  [4,] 0.0231536159 0.02315362 0.023153616
#>  [5,] 0.1573417650 0.15734176 0.157341765
#>  [6,] 0.1039960212 0.37519549 0.005873268
#>  [7,] 0.0785711583 0.07857116 0.078571158
#>  [8,] 0.2235139603 0.22351396 0.223513960
#>  [9,] 0.2317339168 0.23173392 0.231733917
#> [10,] 0.6824234353 0.12101807 0.682423435
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15791224 0.15791224 0.13570920
#>  [2,] 0.02482933 0.02482933 0.35792010
#>  [3,] 0.25593727 0.01863630 0.01863630
#>  [4,] 0.16244657 0.16244657 0.16244657
#>  [5,] 1.12948996 1.43654542 1.33140675
#>  [6,] 0.02899063 0.02899063 0.02899063
#>  [7,] 0.29797007 0.12896562 0.12789312
#>  [8,] 0.20127928 0.20127928 0.18347358
#>  [9,] 0.15216805 0.15216805 0.15216805
#> [10,] 0.12769867 0.12769867 0.12769867
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]      [,2]        [,3]
#>  [1,] 0.24725123 0.3718556 0.080347336
#>  [2,] 0.09497840 0.0949784 0.002720294
#>  [3,] 0.07471667 0.2428330 0.100833384
#>  [4,] 0.01882914 0.4383350 0.438334990
#>  [5,] 0.05583011 0.1235362 0.059444686
#>  [6,] 0.92009234 0.8448968 0.920092344
#>  [7,] 0.95508272 0.4575801 0.457580103
#>  [8,] 1.00933690 1.0093369 1.059781267
#>  [9,] 0.79714193 0.3369415 1.081186990
#> [10,] 1.23601454 1.2360145 0.273563336
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.384808318 0.384808318 0.384808318
#>  [2,] 0.002380689 0.441264441 0.441264441
#>  [3,] 0.228594376 0.700162265 0.228594376
#>  [4,] 0.016528876 0.016528876 0.016528876
#>  [5,] 0.375459795 1.020404401 1.486715554
#>  [6,] 1.524738273 3.085518881 0.611354844
#>  [7,] 0.693644340 0.283811344 0.289996505
#>  [8,] 0.028108777 1.079460937 1.860854088
#>  [9,] 0.475854999 0.475854999 0.475854999
#> [10,] 0.007715799 0.007715799 0.007715799

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.57264721 1.59161583 0.30898219
#>  [2,] 0.23134174 0.90749906 0.98077720
#>  [3,] 0.15693634 0.03908895 0.15693634
#>  [4,] 0.23158880 0.23158880 0.19008436
#>  [5,] 0.58513501 0.58513501 0.07457280
#>  [6,] 0.10690955 0.34350514 0.34350514
#>  [7,] 0.30360417 0.37919802 0.37919802
#>  [8,] 0.76243710 0.76243710 0.21682386
#>  [9,] 0.07461383 0.07461383 0.07461383
#> [10,] 0.58841028 0.58841028 2.66899455
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.07647495 0.07647495 0.076474954
#>  [2,] 0.46379286 0.46379286 0.006290866
#>  [3,] 0.54441334 0.62256604 0.271416809
#>  [4,] 0.19378847 1.06596963 0.089372712
#>  [5,] 0.76498676 0.76498676 0.591245775
#>  [6,] 0.05693662 0.37828871 0.378288709
#>  [7,] 0.39580952 0.39580952 0.395809518
#>  [8,] 0.77502763 0.44473832 0.606587078
#>  [9,] 0.16685347 0.16685347 0.030556485
#> [10,] 0.02795105 0.12221908 0.024057326
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 1.01619674 0.870017295 0.49376295
#>  [2,] 0.31070126 0.006117442 0.31070126
#>  [3,] 0.54129990 0.491656029 0.54129990
#>  [4,] 0.17442999 0.817184415 0.09896078
#>  [5,] 1.22102975 0.678647298 0.06117564
#>  [6,] 0.72249222 0.722492220 0.72249222
#>  [7,] 0.14475539 0.386418023 0.10503459
#>  [8,] 0.03640738 0.036407376 0.03640738
#>  [9,] 0.20719623 0.207196226 0.20719623
#> [10,] 0.08611598 0.495250132 0.56531199
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]        [,2]      [,3]
#>  [1,] 0.46557449 0.008007337 0.3456848
#>  [2,] 0.01083607 0.663456745 0.9713972
#>  [3,] 0.98692569 0.348511410 0.1827348
#>  [4,] 0.60783291 0.099490812 1.6650971
#>  [5,] 0.22001947 0.191056047 0.7207335
#>  [6,] 0.16254683 0.478085778 0.3916190
#>  [7,] 0.68217489 0.259864697 0.2598647
#>  [8,] 0.33589695 0.328910904 0.5660299
#>  [9,] 0.11725253 0.117252528 0.1172525
#> [10,] 0.10390398 0.524020572 0.1039040

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58981806 0.21773212 0.02326432
#>  [2,] 0.05804748 0.22336738 0.22336738
#>  [3,] 1.20170799 0.15149479 0.15149479
#>  [4,] 0.21894085 0.21894085 0.21894085
#>  [5,] 0.19633793 0.19633793 0.19633793
#>  [6,] 0.03940449 0.19153178 0.07545602
#>  [7,] 0.35250901 0.35250901 0.30953517
#>  [8,] 0.50772889 0.25991019 0.13775870
#>  [9,] 0.21325358 0.07864004 0.12994042
#> [10,] 1.25778112 0.35726444 1.00439346
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08235938 0.14993133 0.14993133
#>  [2,] 0.08465799 0.08465799 0.08465799
#>  [3,] 0.08498137 0.52460056 0.31207167
#>  [4,] 0.69487082 0.33452548 0.28053535
#>  [5,] 0.11386880 0.02555331 0.11386880
#>  [6,] 0.53009368 0.23746197 0.44347729
#>  [7,] 0.02589714 0.14114492 0.12259941
#>  [8,] 0.18173962 0.18173962 0.08162635
#>  [9,] 0.63199497 0.12894042 0.40471571
#> [10,] 0.29416432 0.29416432 0.29416432
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>              [,1]        [,2]       [,3]
#>  [1,] 0.091617199 0.091617199 0.09161720
#>  [2,] 0.109464650 0.070657461 0.10946465
#>  [3,] 0.538647864 0.117969825 0.34136019
#>  [4,] 0.437469636 0.177094794 0.17709479
#>  [5,] 0.051008422 0.051008422 0.05100842
#>  [6,] 0.007347205 0.271761690 0.07829173
#>  [7,] 0.379468880 0.005946896 0.37946888
#>  [8,] 0.404601793 0.404601793 0.67693339
#>  [9,] 0.263686238 0.709239286 0.26368624
#> [10,] 1.212964918 0.124238238 1.23784020
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]        [,2]        [,3]
#>  [1,] 0.45477586 0.256855179 0.568562495
#>  [2,] 0.30632138 0.060133265 1.264195271
#>  [3,] 0.34924712 0.451510063 0.293510232
#>  [4,] 0.36838525 0.937164980 0.448917020
#>  [5,] 0.15470729 0.940236938 0.523364312
#>  [6,] 0.14540932 0.367406147 0.367406147
#>  [7,] 0.04479077 0.218895285 0.365858142
#>  [8,] 0.57015628 0.573972330 0.489973800
#>  [9,] 0.01510199 0.004285416 0.004285416
#> [10,] 0.17284425 0.126055364 1.160800127

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.14222592 0.14222592 0.142225919
#>  [2,] 0.20388811 0.21535521 0.215355212
#>  [3,] 0.23432816 0.23432816 0.006549634
#>  [4,] 0.19976641 0.82606524 0.608443221
#>  [5,] 0.52102929 0.33395935 0.093655835
#>  [6,] 0.11499020 0.09553907 0.114990201
#>  [7,] 0.05660322 0.14213531 0.056603216
#>  [8,] 0.08825486 0.77677529 0.833046316
#>  [9,] 0.15940656 0.15940656 0.159406556
#> [10,] 0.39223214 0.26620238 0.058717763
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07089303 0.07089303 0.07089303
#>  [2,] 0.30856484 0.39173897 0.30856484
#>  [3,] 0.87751424 0.05170645 0.10937014
#>  [4,] 0.49988388 2.73074713 0.73331561
#>  [5,] 0.11999816 0.11999816 0.45940650
#>  [6,] 0.04629597 0.04629597 0.04629597
#>  [7,] 0.08883426 0.03400333 0.62879268
#>  [8,] 0.16290289 2.09934473 0.16290289
#>  [9,] 0.63991047 0.12942364 0.12942364
#> [10,] 0.29554555 0.11904262 0.29554555
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.389316828 0.219326640 0.389316828
#>  [2,] 0.013606694 0.013606694 0.235038296
#>  [3,] 0.194421185 0.616750377 0.194421185
#>  [4,] 0.137814700 0.638739176 0.137814700
#>  [5,] 0.039805398 0.018523785 0.440930939
#>  [6,] 1.242763898 0.089873925 0.089873925
#>  [7,] 0.304460845 0.206356768 0.304460845
#>  [8,] 0.000295219 0.000295219 0.000295219
#>  [9,] 0.557387649 0.092520300 0.161698221
#> [10,] 0.127347205 0.186240915 0.589987725
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.07744806 0.07744806 0.1111095
#>  [2,] 0.07072011 0.56465228 0.1858718
#>  [3,] 0.02883653 0.15046118 0.2989314
#>  [4,] 0.86404353 0.20485167 0.7573549
#>  [5,] 0.79445129 0.25028805 0.9282646
#>  [6,] 0.18278037 0.68740439 0.1827804
#>  [7,] 0.11870254 0.02967473 0.1187025
#>  [8,] 0.71594192 0.28385256 0.4945078
#>  [9,] 0.68532459 0.44622576 0.4996856
#> [10,] 0.66718183 0.16074217 0.4357176