Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43944582 0.06200322 1.06248235
#>  [2,] 1.29038719 0.76793544 3.57588745
#>  [3,] 1.24709667 0.58248526 1.24709667
#>  [4,] 0.01763706 0.01763706 0.01763706
#>  [5,] 1.59984889 1.36983662 1.59984889
#>  [6,] 2.06076119 0.37704680 0.01695955
#>  [7,] 0.56061267 0.34576859 1.19831242
#>  [8,] 0.73353558 1.13428049 1.69204798
#>  [9,] 0.09055925 0.42733801 0.91665516
#> [10,] 0.49308323 0.14782133 1.50768928
## comonotone
rpextmo(10, 3, a = 1)
#>              [,1]        [,2]        [,3]
#>  [1,] 2.170878288 2.170878288 2.170878288
#>  [2,] 0.594375195 0.594375195 0.594375195
#>  [3,] 0.908744062 0.908744062 0.908744062
#>  [4,] 0.180366381 0.180366381 0.180366381
#>  [5,] 2.328656116 2.328656116 2.328656116
#>  [6,] 0.003196365 0.003196365 0.003196365
#>  [7,] 1.884587461 1.884587461 1.884587461
#>  [8,] 0.010053226 0.010053226 0.010053226
#>  [9,] 0.959863831 0.959863831 0.959863831
#> [10,] 0.303678223 0.303678223 0.303678223
## independence
rpextmo(10, 3, b = 1)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.16405380 2.211126359 0.38623268
#>  [2,] 1.17983444 2.309922578 1.48536679
#>  [3,] 1.40511154 0.107589393 0.71280477
#>  [4,] 0.88969842 3.507973942 0.35060270
#>  [5,] 1.78528514 0.005387444 2.41184217
#>  [6,] 0.03699807 0.834543519 0.06138128
#>  [7,] 0.48058293 4.107733687 0.13172404
#>  [8,] 0.42969368 0.581596993 1.83750648
#>  [9,] 0.18249623 0.852440827 0.02759086
#> [10,] 0.23571135 1.104064615 0.96243800

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3755880 0.4193688 0.8891496
#>  [2,] 0.6420015 0.4385483 1.3849610
#>  [3,] 1.3978880 1.3366450 1.9855189
#>  [4,] 0.5085071 0.7161952 0.9972828
#>  [5,] 0.4137510 0.4137510 0.4032345
#>  [6,] 1.0417029 5.3374199 0.9997527
#>  [7,] 0.4935950 3.1155647 1.6059710
#>  [8,] 0.4429002 2.3324849 0.8914533
#>  [9,] 2.5090909 1.2294816 1.0770069
#> [10,] 0.8294696 0.8294696 0.3649919
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7887558 0.7887558 0.7887558
#>  [2,] 0.1322166 0.1322166 0.1322166
#>  [3,] 0.8387427 0.8387427 0.8387427
#>  [4,] 0.4196742 0.4196742 0.4196742
#>  [5,] 0.5609576 0.5609576 0.5609576
#>  [6,] 2.9434102 2.9434102 2.9434102
#>  [7,] 0.3675554 0.3675554 0.3675554
#>  [8,] 0.6352486 0.6352486 0.6352486
#>  [9,] 0.4240581 0.4240581 0.4240581
#> [10,] 1.6270595 1.6270595 1.6270595
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>              [,1]         [,2]       [,3]
#>  [1,] 4.142181461 0.2549584349 1.79768701
#>  [2,] 1.665590217 0.3531743367 0.74775103
#>  [3,] 0.777809503 0.0002668165 4.54373007
#>  [4,] 1.364710811 0.3957187960 1.67831667
#>  [5,] 0.849294268 0.4595614495 1.78063504
#>  [6,] 1.161012772 1.0216001803 0.13120887
#>  [7,] 0.538270204 2.6105168665 0.27121024
#>  [8,] 0.001805052 2.7583144921 0.03331008
#>  [9,] 1.571020180 0.4245218472 0.40392940
#> [10,] 0.137521453 0.8155695247 3.24509398

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.94094208 0.62054578 3.32951977
#>  [2,] 2.39465753 1.65581155 1.67168430
#>  [3,] 0.41572253 2.46397987 2.90104832
#>  [4,] 3.74983563 2.16603740 1.70256179
#>  [5,] 0.09389364 0.09389364 0.09389364
#>  [6,] 0.82904095 3.88217588 1.80731469
#>  [7,] 1.00696088 0.78499593 0.40444438
#>  [8,] 0.22138401 0.50771956 0.50771956
#>  [9,] 3.07776543 1.75358533 0.22543163
#> [10,] 8.06798450 2.52401144 6.43265749
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13858655 0.13858655 0.13858655
#>  [2,] 2.71469880 2.71469880 2.71469880
#>  [3,] 0.20741507 0.20741507 0.20741507
#>  [4,] 0.11225969 0.11225969 0.11225969
#>  [5,] 1.92593173 1.92593173 1.92593173
#>  [6,] 1.32686344 1.32686344 1.32686344
#>  [7,] 0.22425526 0.22425526 0.22425526
#>  [8,] 1.85494777 1.85494777 1.85494777
#>  [9,] 0.56428629 0.56428629 0.56428629
#> [10,] 0.04506891 0.04506891 0.04506891
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.68190627 1.9548850 0.1024855
#>  [2,] 0.28855017 1.0729556 0.4971823
#>  [3,] 0.05214345 0.1957367 0.3132046
#>  [4,] 0.28548205 2.4523992 0.2054066
#>  [5,] 0.63568442 0.3028270 0.7767042
#>  [6,] 4.00178563 0.4713698 0.3894317
#>  [7,] 0.31955509 2.2697410 2.0925751
#>  [8,] 0.27315550 0.3321665 1.5983734
#>  [9,] 0.28054549 0.8563496 1.9777774
#> [10,] 2.29563356 1.8729553 0.1429578

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.51461242 1.5749389 0.1512669
#>  [2,] 0.83767592 0.6430563 2.0655485
#>  [3,] 0.03483864 1.4974202 0.1096407
#>  [4,] 1.00437250 0.7772226 4.1813778
#>  [5,] 1.80668675 0.4964714 0.1254137
#>  [6,] 0.86318512 0.7949354 1.3606599
#>  [7,] 3.63390045 5.1427754 0.1425999
#>  [8,] 2.31712702 0.7885724 0.3077530
#>  [9,] 0.02492519 0.9222626 0.3738542
#> [10,] 0.45391497 1.6308719 5.7476718
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5879534 0.5879534 0.5879534
#>  [2,] 0.8049404 0.8049404 0.8049404
#>  [3,] 1.8231112 1.8231112 1.8231112
#>  [4,] 2.1109433 2.1109433 2.1109433
#>  [5,] 1.2365898 1.2365898 1.2365898
#>  [6,] 0.4103797 0.4103797 0.4103797
#>  [7,] 2.9198117 2.9198117 2.9198117
#>  [8,] 1.4196162 1.4196162 1.4196162
#>  [9,] 2.1280780 2.1280780 2.1280780
#> [10,] 0.1029995 0.1029995 0.1029995
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7330475 0.34589811 1.0658607
#>  [2,] 0.3566889 0.31727761 0.8053135
#>  [3,] 2.4529378 1.54068203 0.5526553
#>  [4,] 1.8835176 0.77271793 1.4767502
#>  [5,] 3.2202087 0.65994482 0.2633776
#>  [6,] 1.6173033 0.12337956 2.6387367
#>  [7,] 0.6951390 0.10157272 2.1694532
#>  [8,] 0.9019262 0.09677449 1.5719249
#>  [9,] 1.2362187 0.06011506 0.2468004
#> [10,] 0.6419937 0.61227799 0.2317497

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]        [,3]
#>  [1,] 1.1620446 0.7626852 1.162044600
#>  [2,] 0.3351537 2.2336652 0.003952763
#>  [3,] 0.4302820 0.5393560 0.517238977
#>  [4,] 3.4776046 3.5173911 0.903985199
#>  [5,] 1.7520514 0.8110021 1.721806957
#>  [6,] 0.4607593 0.1263632 0.458340316
#>  [7,] 0.1224745 1.1028339 0.385531612
#>  [8,] 0.3042540 1.6887887 0.557373695
#>  [9,] 0.1328177 4.1365328 2.725285405
#> [10,] 0.5842145 1.6063044 1.606304380
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24677919 0.24677919 0.24677919
#>  [2,] 0.29172126 0.29172126 0.29172126
#>  [3,] 2.37081815 2.37081815 2.37081815
#>  [4,] 1.47905132 1.47905132 1.47905132
#>  [5,] 2.52592399 2.52592399 2.52592399
#>  [6,] 0.54052742 0.54052742 0.54052742
#>  [7,] 0.51652812 0.51652812 0.51652812
#>  [8,] 2.14944082 2.14944082 2.14944082
#>  [9,] 0.19397292 0.19397292 0.19397292
#> [10,] 0.01315397 0.01315397 0.01315397
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]      [,2]        [,3]
#>  [1,] 0.21916540 0.2539909 2.044550924
#>  [2,] 0.37858191 0.7916488 2.598791882
#>  [3,] 2.61292472 0.3892970 0.001631889
#>  [4,] 0.55010171 0.8723387 0.310670028
#>  [5,] 2.14019844 0.1724573 0.568029121
#>  [6,] 1.49343730 1.8553001 0.529998418
#>  [7,] 0.92186921 0.9162575 0.160380115
#>  [8,] 1.31188035 1.1833280 0.342962255
#>  [9,] 0.91270856 0.2429720 1.304714310
#> [10,] 0.02375184 2.2800589 0.010211607

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.37602054 0.3450125 0.3760205
#>  [2,] 3.10457643 1.0436361 0.3498109
#>  [3,] 0.59939628 0.5993963 0.2532042
#>  [4,] 0.76547719 4.0292780 0.1348790
#>  [5,] 0.66968910 0.6696891 0.5437196
#>  [6,] 0.91386492 4.2145603 0.1200726
#>  [7,] 0.18876021 0.9214367 4.8578087
#>  [8,] 0.04007252 0.6386814 1.0098823
#>  [9,] 1.20435365 1.0013289 0.5027177
#> [10,] 0.87046157 0.6345826 0.6285804
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7208449 0.7208449 0.7208449
#>  [2,] 0.5557618 0.5557618 0.5557618
#>  [3,] 0.1439258 0.1439258 0.1439258
#>  [4,] 1.0903652 1.0903652 1.0903652
#>  [5,] 2.5517196 2.5517196 2.5517196
#>  [6,] 1.9483096 1.9483096 1.9483096
#>  [7,] 1.2627808 1.2627808 1.2627808
#>  [8,] 2.0270325 2.0270325 2.0270325
#>  [9,] 0.5075756 0.5075756 0.5075756
#> [10,] 1.4158095 1.4158095 1.4158095
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]        [,2]       [,3]
#>  [1,] 2.03201465 0.278779380 2.43568525
#>  [2,] 0.01616589 3.623062668 3.95658802
#>  [3,] 1.57256626 0.868264131 0.35283575
#>  [4,] 0.06146502 1.187136686 2.16662609
#>  [5,] 0.87827834 0.487769030 0.56738015
#>  [6,] 0.47781832 2.149979662 0.03395222
#>  [7,] 0.17325632 0.002162647 0.93148399
#>  [8,] 1.63235580 0.011315665 1.59355418
#>  [9,] 1.55635324 0.465696793 2.33749904
#> [10,] 0.24386562 1.023907195 3.76855520

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>              [,1]      [,2]       [,3]
#>  [1,] 1.861008895 0.5049711 2.11889420
#>  [2,] 0.379528045 2.9052805 2.90528046
#>  [3,] 0.007212557 1.5664068 0.09860913
#>  [4,] 1.146069081 1.8819800 0.62417810
#>  [5,] 0.649624843 2.0326155 2.03261549
#>  [6,] 3.087755173 1.9745717 5.77934325
#>  [7,] 2.648994043 1.5323674 3.95632982
#>  [8,] 0.886518796 2.7646095 0.35207593
#>  [9,] 1.133071423 1.1462840 1.14628403
#> [10,] 1.428601763 1.4561407 2.81744556
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.43688599 0.43688599 0.43688599
#>  [2,] 0.25850054 0.25850054 0.25850054
#>  [3,] 1.58448773 1.58448773 1.58448773
#>  [4,] 2.02415449 2.02415449 2.02415449
#>  [5,] 1.57686714 1.57686714 1.57686714
#>  [6,] 0.74950277 0.74950277 0.74950277
#>  [7,] 0.01847786 0.01847786 0.01847786
#>  [8,] 0.33240431 0.33240431 0.33240431
#>  [9,] 0.50627624 0.50627624 0.50627624
#> [10,] 1.19644308 1.19644308 1.19644308
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>              [,1]         [,2]      [,3]
#>  [1,] 0.085998558 2.156576e+00 1.6672375
#>  [2,] 0.009917178 1.680414e+00 1.3612383
#>  [3,] 1.700841701 1.577155e+00 0.3478628
#>  [4,] 0.371664799 1.327576e-01 3.1897524
#>  [5,] 0.979879993 1.003839e+00 5.7038472
#>  [6,] 0.188934898 9.417953e-01 0.1834343
#>  [7,] 0.750644960 7.187967e-06 0.4060973
#>  [8,] 0.799939508 1.407830e+00 2.3752191
#>  [9,] 0.769430208 2.393838e-01 1.1256323
#> [10,] 0.142014888 4.504562e-01 0.3154922

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 2.21470403 1.3050977 1.87936017
#>  [2,] 0.02678631 2.1721772 2.42053504
#>  [3,] 1.91596263 6.6987479 5.27843842
#>  [4,] 5.49190859 0.6859418 0.84999530
#>  [5,] 0.33354963 0.3115713 0.33354963
#>  [6,] 2.36904207 2.3690421 1.04617057
#>  [7,] 3.36040918 0.6824730 0.67156953
#>  [8,] 0.76045158 1.7999030 0.01602885
#>  [9,] 0.47298013 0.6695272 0.66952722
#> [10,] 2.59721562 0.1301144 3.07309037
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1021170 1.1021170 1.1021170
#>  [2,] 0.3994631 0.3994631 0.3994631
#>  [3,] 3.1236807 3.1236807 3.1236807
#>  [4,] 1.9989118 1.9989118 1.9989118
#>  [5,] 1.7114060 1.7114060 1.7114060
#>  [6,] 0.1491573 0.1491573 0.1491573
#>  [7,] 1.3815800 1.3815800 1.3815800
#>  [8,] 0.1243818 0.1243818 0.1243818
#>  [9,] 0.3039177 0.3039177 0.3039177
#> [10,] 0.8653374 0.8653374 0.8653374
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.19196028 0.441488541 1.23968133
#>  [2,] 0.27381638 0.721284352 0.04337208
#>  [3,] 2.58109378 0.144698947 1.09585136
#>  [4,] 0.06879198 4.406657058 2.16829505
#>  [5,] 0.71938900 0.375857239 1.19426505
#>  [6,] 0.19597536 0.771835286 0.33267205
#>  [7,] 0.89293618 0.313228838 2.95828730
#>  [8,] 0.31761224 1.117550121 0.42236392
#>  [9,] 2.52666597 0.611847828 0.51298817
#> [10,] 1.39267464 0.003257502 0.13853281

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]     [,2]      [,3]
#>  [1,] 3.4507547 3.450755 0.4530077
#>  [2,] 2.4980264 2.799048 2.7990476
#>  [3,] 2.4373529 1.431373 0.3048605
#>  [4,] 1.5449214 3.244676 3.2446757
#>  [5,] 0.6425561 1.581386 0.5370988
#>  [6,] 1.4106886 2.827317 3.6513016
#>  [7,] 0.5618741 2.986845 2.8040772
#>  [8,] 0.5673824 1.301805 3.1387035
#>  [9,] 0.1845883 2.021865 3.0430872
#> [10,] 1.8576902 1.136161 4.7883786
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31578603 0.31578603 0.31578603
#>  [2,] 0.81943934 0.81943934 0.81943934
#>  [3,] 1.79005111 1.79005111 1.79005111
#>  [4,] 0.12860535 0.12860535 0.12860535
#>  [5,] 0.54760335 0.54760335 0.54760335
#>  [6,] 0.12132760 0.12132760 0.12132760
#>  [7,] 3.64931906 3.64931906 3.64931906
#>  [8,] 0.18699885 0.18699885 0.18699885
#>  [9,] 0.08532985 0.08532985 0.08532985
#> [10,] 0.33739929 0.33739929 0.33739929
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.89824584 1.2431670 2.66499287
#>  [2,] 1.16083986 0.4088811 1.69308322
#>  [3,] 1.00938623 0.2416933 0.99116780
#>  [4,] 1.92146965 0.6613850 0.82611013
#>  [5,] 0.09628188 0.4438910 2.05912154
#>  [6,] 0.39871447 1.1032953 1.99834061
#>  [7,] 2.29513506 1.6993774 0.09413799
#>  [8,] 0.28520445 0.7412533 2.45392001
#>  [9,] 0.75020616 0.6692981 1.25537304
#> [10,] 0.19564556 0.5691404 1.14831194

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.5516212 1.22723409 1.55162115
#>  [2,] 0.2606615 0.35193139 0.62009942
#>  [3,] 0.0715201 3.22554897 0.89112636
#>  [4,] 1.1746628 4.35303772 0.40107649
#>  [5,] 0.8305663 4.28883165 3.60555402
#>  [6,] 5.3613252 4.97812517 1.45831208
#>  [7,] 0.3837549 3.02477189 0.01138601
#>  [8,] 0.6683516 0.28595215 0.69869234
#>  [9,] 0.4831069 1.24379767 1.46179263
#> [10,] 1.4195519 0.02281153 0.64520414
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2503541 1.2503541 1.2503541
#>  [2,] 2.1998457 2.1998457 2.1998457
#>  [3,] 2.4266213 2.4266213 2.4266213
#>  [4,] 1.8706514 1.8706514 1.8706514
#>  [5,] 0.8578449 0.8578449 0.8578449
#>  [6,] 1.3214817 1.3214817 1.3214817
#>  [7,] 0.6845171 0.6845171 0.6845171
#>  [8,] 3.2596737 3.2596737 3.2596737
#>  [9,] 2.3128246 2.3128246 2.3128246
#> [10,] 0.7405454 0.7405454 0.7405454
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.66820654 0.79560103 0.6282987
#>  [2,] 2.23902544 0.58842690 0.3878730
#>  [3,] 1.45045973 2.61630775 1.6974997
#>  [4,] 0.78048796 1.39964421 1.1171925
#>  [5,] 1.06709032 0.03506524 0.8040791
#>  [6,] 1.00667183 0.05623105 0.5256042
#>  [7,] 0.02139251 0.27966008 0.1249470
#>  [8,] 0.14554070 0.20216868 0.9781200
#>  [9,] 0.77507485 0.00744630 0.9823238
#> [10,] 2.37083327 1.06710676 1.7534952

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.67700703 0.66828261 2.17330422
#>  [2,] 0.77368815 0.29962055 0.77368815
#>  [3,] 0.05566319 0.05566319 0.05566319
#>  [4,] 0.08777763 0.08777763 0.08777763
#>  [5,] 0.66508286 0.67692906 0.07332068
#>  [6,] 0.54055813 1.79897299 0.54055813
#>  [7,] 0.25017344 0.32757471 0.91616841
#>  [8,] 1.93010397 2.76199806 0.98919862
#>  [9,] 1.01698042 0.56463560 1.01698042
#> [10,] 0.32639501 1.37516182 1.02923470
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39503420 0.21241394 1.03114237
#>  [2,] 0.10976777 0.08808772 0.08808772
#>  [3,] 0.69577470 0.02621417 0.02621417
#>  [4,] 0.06001584 0.38657221 0.16463038
#>  [5,] 0.09430641 0.69043042 0.43972092
#>  [6,] 1.06690219 1.37005600 0.21616043
#>  [7,] 0.14000961 0.14000961 0.70522108
#>  [8,] 0.33428396 0.77478448 0.64567894
#>  [9,] 0.08335095 0.42307342 0.12172592
#> [10,] 0.44821609 0.44821609 0.46957248
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.90065997 0.90065997 0.1104613
#>  [2,] 0.20883893 1.08858961 1.0207941
#>  [3,] 0.58686037 0.58686037 0.5868604
#>  [4,] 0.27958470 0.23346434 0.2795847
#>  [5,] 0.34368952 1.32995215 0.1520825
#>  [6,] 0.13413170 0.16288548 0.1628855
#>  [7,] 1.33924223 0.05452765 0.6954092
#>  [8,] 0.55989010 0.19255319 0.5111579
#>  [9,] 0.09675865 0.14359275 0.1435927
#> [10,] 1.31873082 1.61125884 0.8635346
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5086658 0.7781322 1.2552095
#>  [2,] 0.6391690 0.5022840 0.6391690
#>  [3,] 0.2446277 0.1495996 0.2446277
#>  [4,] 0.2749995 0.1690656 0.2749995
#>  [5,] 0.1270088 0.1270088 0.1270088
#>  [6,] 0.1270654 0.9648009 0.1270654
#>  [7,] 0.3940136 2.0329478 0.3940136
#>  [8,] 0.5602775 0.1896421 0.5602775
#>  [9,] 0.2170749 0.1596701 1.7894502
#> [10,] 0.7386720 1.3377036 0.7386720
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.15047358 0.43404251 1.46244747
#>  [2,] 0.17710900 0.88693830 0.88693830
#>  [3,] 0.41620280 0.30085319 0.34142490
#>  [4,] 3.13215500 1.44578904 1.32539184
#>  [5,] 0.56607844 0.97127800 1.41289934
#>  [6,] 0.87056381 0.68840509 0.65993752
#>  [7,] 0.37670318 0.18634618 0.18634618
#>  [8,] 0.03300763 0.03585430 0.58113325
#>  [9,] 0.11614349 0.08851863 0.08851863
#> [10,] 0.35119613 0.74355993 0.74355993

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]       [,2]      [,3]
#>  [1,] 7.18353265 4.10426592 5.1223753
#>  [2,] 0.53543672 2.07614239 1.0992312
#>  [3,] 0.15984850 0.15984850 0.1449069
#>  [4,] 1.08180632 1.08180632 1.0818063
#>  [5,] 0.37832939 2.27331892 1.7879575
#>  [6,] 0.02793756 0.26599914 0.2659991
#>  [7,] 0.55709019 0.26155219 0.2443018
#>  [8,] 1.84600320 1.84600320 1.8460032
#>  [9,] 0.27088410 0.51998049 1.1706164
#> [10,] 0.73276208 0.06972825 0.0734284
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.63709724 0.5478893 1.6370972
#>  [2,] 1.45055710 9.4317847 2.9579725
#>  [3,] 2.02140082 2.0821808 3.6743091
#>  [4,] 2.96054966 2.9181066 0.2503275
#>  [5,] 0.51802759 0.2831482 1.0739724
#>  [6,] 5.83342798 0.4201709 7.3976480
#>  [7,] 3.78657309 7.1295654 2.2803110
#>  [8,] 0.08231181 1.9408803 2.5842968
#>  [9,] 0.71763156 0.5787290 0.2312831
#> [10,] 0.41469772 0.8196402 1.0363768
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.85835898 0.70086358 0.9017676
#>  [2,] 0.39033142 3.31491235 0.7592818
#>  [3,] 3.56746252 1.49856476 1.0612117
#>  [4,] 1.08990825 1.37739527 0.4884021
#>  [5,] 0.24979020 0.38280008 3.2969295
#>  [6,] 0.47267091 0.12787073 0.7527486
#>  [7,] 0.60505893 0.40032022 0.6050589
#>  [8,] 0.67686030 1.36579424 0.6836372
#>  [9,] 0.08787498 1.19450315 1.1945031
#> [10,] 1.11234133 0.05528461 3.7434227
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.86917603 0.43943150 0.50076119
#>  [2,] 1.57622067 1.08719380 1.85983687
#>  [3,] 0.46025355 0.46025355 0.46025355
#>  [4,] 1.22442837 0.82174020 2.05973782
#>  [5,] 0.79789925 3.61313230 0.42584212
#>  [6,] 0.16887185 0.57016756 0.24497357
#>  [7,] 0.54129232 0.54129232 0.54129232
#>  [8,] 2.35361345 0.03118306 0.03171521
#>  [9,] 1.22736812 0.70393709 1.86633581
#> [10,] 0.06132064 0.06132064 0.06132064
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.63283496 2.01401639 2.63283496
#>  [2,] 0.43183424 0.72362132 0.34107828
#>  [3,] 6.06028255 1.38087925 1.36732603
#>  [4,] 3.09817036 0.69435669 0.78847301
#>  [5,] 3.90554239 0.57511986 0.13909120
#>  [6,] 2.77822126 1.59575691 2.05986579
#>  [7,] 2.26559606 1.24282725 2.08544002
#>  [8,] 0.50934063 0.50934063 0.50934063
#>  [9,] 0.31994913 0.95100718 2.11425379
#> [10,] 0.01777987 0.01777987 0.01777987

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>              [,1]      [,2]       [,3]
#>  [1,] 0.689449681 0.6894497 0.68944968
#>  [2,] 0.308791110 0.1638460 0.15102928
#>  [3,] 0.152793517 1.3580158 0.46928703
#>  [4,] 1.613368677 1.0978192 1.83439039
#>  [5,] 0.098352698 0.0983527 0.20337853
#>  [6,] 0.537451487 0.5508582 0.71525561
#>  [7,] 1.014633323 1.6547944 1.01463332
#>  [8,] 0.392199618 0.3921996 0.38446625
#>  [9,] 0.004401911 0.4422943 0.44229433
#> [10,] 0.058189751 0.9037771 0.05818975
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11185348 0.11185348 0.11185348
#>  [2,] 0.10050808 0.92621884 0.94989046
#>  [3,] 0.05650551 0.24439445 0.05650551
#>  [4,] 1.96742180 1.96742180 0.24725398
#>  [5,] 1.08682528 1.08753883 1.08753883
#>  [6,] 0.80846365 0.80846365 0.80846365
#>  [7,] 0.06704921 0.03190810 0.06704921
#>  [8,] 0.72722459 1.42347913 0.81872687
#>  [9,] 0.06052911 0.06052911 0.06052911
#> [10,] 0.14698828 0.14698828 0.14698828
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3047992 0.3047992 0.9607593
#>  [2,] 3.4561713 2.5581266 0.2769528
#>  [3,] 0.3741915 0.1231462 0.2529749
#>  [4,] 0.3309250 0.4663057 0.4663057
#>  [5,] 0.1406751 0.1406751 0.1406751
#>  [6,] 0.1295659 0.1295659 0.1334641
#>  [7,] 1.3079279 1.3079279 0.1653669
#>  [8,] 1.9407765 0.6061284 0.7998357
#>  [9,] 1.3242048 1.4272742 1.3242048
#> [10,] 0.1263806 0.1195482 0.1263806
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 1.421438761 0.57067651 0.01265015
#>  [2,] 0.206223484 0.20622348 0.20622348
#>  [3,] 0.005982941 0.23091247 0.23091247
#>  [4,] 0.006222797 0.17063996 1.08693786
#>  [5,] 0.066175563 0.06617556 0.06617556
#>  [6,] 0.804710359 0.47500734 0.80471036
#>  [7,] 0.324950919 0.47428745 0.47428745
#>  [8,] 0.308025582 0.30802558 0.25384232
#>  [9,] 1.755690845 1.75569084 1.75569084
#> [10,] 0.116374537 0.11637454 0.11637454
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02405121 0.02405121 0.02405121
#>  [2,] 0.41136991 0.41136991 0.41136991
#>  [3,] 1.34430756 3.49693172 0.86903705
#>  [4,] 0.07135937 0.07135937 0.07135937
#>  [5,] 0.87658120 0.87658120 0.87658120
#>  [6,] 2.33108956 0.60819352 1.15383316
#>  [7,] 0.04828398 0.07988116 0.07988116
#>  [8,] 1.03047178 2.12142264 0.01611951
#>  [9,] 1.23817129 1.07182897 0.90392553
#> [10,] 0.49234082 1.12230244 0.89155637

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01579594 0.01579594 0.01579594
#>  [2,] 0.02008763 0.68650498 0.02008763
#>  [3,] 0.12601999 0.12601999 0.12601999
#>  [4,] 0.25460225 0.47975221 0.01062168
#>  [5,] 0.08017957 0.05893751 0.08017957
#>  [6,] 0.25891107 0.32998617 0.25891107
#>  [7,] 0.05673417 0.24367001 0.24367001
#>  [8,] 0.18342487 0.47082347 0.21847876
#>  [9,] 0.14480945 0.15485140 0.15485140
#> [10,] 0.33844398 0.18421725 0.34773755
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41963878 0.78045376 1.07893818
#>  [2,] 0.13202919 0.13202919 0.13202919
#>  [3,] 0.73602331 0.37480683 0.07719787
#>  [4,] 1.03519797 1.05695877 1.21773203
#>  [5,] 0.60004886 0.03263295 0.03263295
#>  [6,] 0.44092517 0.13631533 0.46831001
#>  [7,] 0.03335332 0.03335332 0.03335332
#>  [8,] 0.18324487 0.53770532 0.53770532
#>  [9,] 0.22539441 0.83499700 0.36753027
#> [10,] 0.38253904 0.05035008 0.23589952
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60762828 0.44084543 1.28558063
#>  [2,] 0.33422433 0.33422433 0.11907552
#>  [3,] 1.17375826 0.43818585 0.16322799
#>  [4,] 0.32276789 0.46619417 0.22381462
#>  [5,] 0.11730539 0.11730539 0.11730539
#>  [6,] 0.29917835 0.06341391 0.29917835
#>  [7,] 1.43125548 0.18718029 1.28593673
#>  [8,] 0.09565133 0.23176141 0.09565133
#>  [9,] 0.03932267 0.03932267 0.03932267
#> [10,] 0.34546480 0.24404963 0.34546480
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.48851145 0.46925939 0.22217434
#>  [2,] 0.73307835 0.62904328 0.27880336
#>  [3,] 0.67017187 0.64205164 0.47275448
#>  [4,] 1.59667974 0.12826519 0.16219817
#>  [5,] 0.34222956 1.12207980 1.04555063
#>  [6,] 0.05994958 0.05994958 0.85252882
#>  [7,] 0.07236932 0.07236932 0.07236932
#>  [8,] 0.34269069 0.38610861 0.11411814
#>  [9,] 1.84484380 0.40125887 0.27312863
#> [10,] 0.17359060 0.45176300 0.72503207

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.57541032 0.9251886 0.2881619
#>  [2,] 0.32494903 0.6499853 0.4882475
#>  [3,] 0.37800642 0.1146101 0.1146101
#>  [4,] 0.12043955 0.1204395 0.1204395
#>  [5,] 0.05251235 0.1424468 0.1424468
#>  [6,] 0.21932097 0.9948534 0.7072850
#>  [7,] 0.23434927 0.1045198 0.4606940
#>  [8,] 0.35189522 0.3943684 0.4042113
#>  [9,] 0.70623390 0.2172063 0.7652244
#> [10,] 0.02895111 0.1130184 0.2155761
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31823218 0.09348580 0.13832716
#>  [2,] 0.37560867 0.24863619 0.24863619
#>  [3,] 0.25861365 0.71331809 0.84592356
#>  [4,] 0.28513374 0.21653876 0.01929898
#>  [5,] 0.08494725 0.04257964 0.05116550
#>  [6,] 0.07858877 0.33155517 0.10394909
#>  [7,] 1.19009567 0.44492025 1.50056236
#>  [8,] 0.36324049 0.36324049 0.84958737
#>  [9,] 0.01247883 0.08180933 0.21725082
#> [10,] 0.13794380 0.24160091 0.59811033
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.09058340 0.26724736 0.1761259
#>  [2,] 0.90918390 1.04440943 0.2790242
#>  [3,] 0.79885016 0.71841965 0.2381324
#>  [4,] 0.15467559 0.07679725 0.1224847
#>  [5,] 0.11636223 0.89436181 0.4494321
#>  [6,] 0.12231948 0.12782919 0.2947291
#>  [7,] 0.27075512 0.29969598 0.8175659
#>  [8,] 0.01818298 0.42373704 0.3505429
#>  [9,] 0.48575171 0.06518555 0.2313837
#> [10,] 0.39508510 0.14045722 0.4311550
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2051926 1.1670787 1.04861482
#>  [2,] 0.5353742 0.2878777 0.06339345
#>  [3,] 0.5284921 0.5280442 0.52804421
#>  [4,] 0.1607166 0.5442278 0.41323540
#>  [5,] 0.1640748 0.1401883 1.66517511
#>  [6,] 0.4836601 0.6674509 1.71574375
#>  [7,] 0.1955892 0.4935813 0.19558922
#>  [8,] 0.1959464 0.1959464 0.19594644
#>  [9,] 0.3788770 0.3520341 0.35203408
#> [10,] 0.3542503 0.2480884 1.56980901

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19586313 0.21560060 0.53227060
#>  [2,] 0.05357579 0.05357579 0.32940374
#>  [3,] 0.20928765 0.20928765 0.65089039
#>  [4,] 0.15264632 0.38612561 0.29339889
#>  [5,] 0.32566042 0.46142756 0.46142756
#>  [6,] 0.30482667 0.30482667 0.30482667
#>  [7,] 0.53506936 0.53506936 0.53506936
#>  [8,] 0.32295002 1.48346182 0.32607465
#>  [9,] 0.23247239 0.15279757 0.15279757
#> [10,] 0.08496547 1.12040712 0.08496547
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13861981 1.38328916 0.13861981
#>  [2,] 0.05810963 0.15221775 0.58818541
#>  [3,] 0.07843447 0.07843447 0.07843447
#>  [4,] 0.11745231 0.11745231 0.02716638
#>  [5,] 0.08504152 0.08504152 0.63367732
#>  [6,] 0.09836579 0.12873212 0.05576858
#>  [7,] 0.08051227 0.16420873 0.16420873
#>  [8,] 0.05692483 2.09515169 0.05692483
#>  [9,] 0.00631536 0.37304234 0.16255298
#> [10,] 0.09729470 0.03020000 0.08029254
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1487065 0.1487065 0.04403462
#>  [2,] 0.3731684 0.0259425 0.54710519
#>  [3,] 0.1563572 0.2009036 0.33430825
#>  [4,] 0.5132665 0.7461797 0.12233194
#>  [5,] 0.5331793 0.5331793 0.53317933
#>  [6,] 1.4119683 1.6839575 0.90973979
#>  [7,] 0.3112423 0.3116014 1.72952255
#>  [8,] 0.4271783 0.8429158 0.74115165
#>  [9,] 0.9883501 0.1581195 0.07690633
#> [10,] 0.2956295 0.3971471 0.48453834
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.44095841 0.44095841 0.203172414
#>  [2,] 0.20824090 0.20824090 0.208240901
#>  [3,] 0.13019565 0.16974575 0.502069508
#>  [4,] 0.04618138 0.04618138 0.046181378
#>  [5,] 1.80054042 0.06138455 0.061384552
#>  [6,] 0.39126709 0.24261833 0.520628384
#>  [7,] 0.26617000 0.30189252 0.112087510
#>  [8,] 0.19729710 0.71108178 0.280097966
#>  [9,] 0.36286919 0.36286919 0.134488234
#> [10,] 0.05664954 0.42217061 0.002850912