Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04330970 4.22103513 4.22103513
#>  [2,] 3.01787740 1.24894204 2.19507161
#>  [3,] 0.01028352 0.01028352 0.01028352
#>  [4,] 0.04388825 4.23998640 3.10113289
#>  [5,] 2.64326728 2.64326728 1.53419469
#>  [6,] 0.24831882 0.08212736 0.66640989
#>  [7,] 0.55833421 0.66374393 0.66374393
#>  [8,] 0.19724618 1.19473930 0.79894398
#>  [9,] 1.87542981 0.21942128 1.13097570
#> [10,] 1.45220579 1.45220579 1.39326610
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14672772 0.14672772 0.14672772
#>  [2,] 0.88933226 0.88933226 0.88933226
#>  [3,] 0.03552019 0.03552019 0.03552019
#>  [4,] 2.62489710 2.62489710 2.62489710
#>  [5,] 2.81168221 2.81168221 2.81168221
#>  [6,] 2.58820363 2.58820363 2.58820363
#>  [7,] 1.00814552 1.00814552 1.00814552
#>  [8,] 0.44177873 0.44177873 0.44177873
#>  [9,] 0.08011792 0.08011792 0.08011792
#> [10,] 0.40494778 0.40494778 0.40494778
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.93559733 1.72128509 0.62226324
#>  [2,] 1.34312067 0.89570685 0.05272101
#>  [3,] 0.61145186 0.20475592 0.80798339
#>  [4,] 2.92088731 0.45222051 0.16989232
#>  [5,] 1.95977367 0.23863507 3.01550075
#>  [6,] 1.34068287 0.86421963 0.27444372
#>  [7,] 3.36648410 0.98254287 0.10868426
#>  [8,] 0.05018691 1.60269287 0.13726707
#>  [9,] 0.77295106 0.04367938 0.53417252
#> [10,] 0.79256430 0.51910776 0.53201732

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3247944 3.38826429 7.0920628
#>  [2,] 3.0907869 0.14922524 1.4543733
#>  [3,] 1.6890045 2.09333342 2.7442805
#>  [4,] 0.2236126 0.09172608 0.9857582
#>  [5,] 1.0194265 0.03421712 1.0194265
#>  [6,] 4.2220632 4.22206316 0.1121140
#>  [7,] 3.6348529 1.10413477 1.5547407
#>  [8,] 3.5425075 2.63213669 0.7439059
#>  [9,] 0.1610406 1.03979278 0.1460146
#> [10,] 0.6583084 0.57971278 0.6583084
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9211550 1.9211550 1.9211550
#>  [2,] 1.1478411 1.1478411 1.1478411
#>  [3,] 1.7383224 1.7383224 1.7383224
#>  [4,] 1.6557383 1.6557383 1.6557383
#>  [5,] 0.7580805 0.7580805 0.7580805
#>  [6,] 0.5055266 0.5055266 0.5055266
#>  [7,] 0.4596066 0.4596066 0.4596066
#>  [8,] 0.5969995 0.5969995 0.5969995
#>  [9,] 0.9979690 0.9979690 0.9979690
#> [10,] 0.9542228 0.9542228 0.9542228
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>            [,1]      [,2]        [,3]
#>  [1,] 1.7118528 0.3892788 0.102427622
#>  [2,] 0.6568169 0.3465741 0.146045628
#>  [3,] 0.8481364 1.0485127 1.178499786
#>  [4,] 0.6885486 4.1257896 1.871374119
#>  [5,] 2.0982335 0.1172374 0.593805592
#>  [6,] 0.1762672 0.2400011 0.007389781
#>  [7,] 0.4422165 0.2637309 1.708342029
#>  [8,] 3.1930048 1.0968620 0.561483782
#>  [9,] 0.3976350 2.6333938 0.673994082
#> [10,] 0.2703435 1.3695619 0.842389073

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.4494656 0.9795408 1.46836523
#>  [2,] 1.1328119 1.1328119 1.13281186
#>  [3,] 0.1005847 0.1005847 0.10058468
#>  [4,] 5.7107391 0.2014545 6.17243730
#>  [5,] 0.8456070 0.1410071 0.97860743
#>  [6,] 0.4659598 0.4659598 0.20367736
#>  [7,] 3.5583425 2.1933504 0.69301472
#>  [8,] 0.8772949 0.8772949 0.06089556
#>  [9,] 1.4616464 3.7725430 2.89429948
#> [10,] 0.5318515 4.5074048 0.68382757
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2226439 0.2226439 0.2226439
#>  [2,] 0.6797905 0.6797905 0.6797905
#>  [3,] 0.5855748 0.5855748 0.5855748
#>  [4,] 0.2788712 0.2788712 0.2788712
#>  [5,] 0.9923336 0.9923336 0.9923336
#>  [6,] 1.9124785 1.9124785 1.9124785
#>  [7,] 0.6262147 0.6262147 0.6262147
#>  [8,] 0.9341514 0.9341514 0.9341514
#>  [9,] 1.1795307 1.1795307 1.1795307
#> [10,] 0.2852012 0.2852012 0.2852012
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0837335 0.1875877 0.7439135
#>  [2,] 2.2022752 0.5873881 0.2627507
#>  [3,] 0.8010803 3.5480660 2.0525068
#>  [4,] 0.1584992 0.2877199 2.9104388
#>  [5,] 0.1334763 0.1806446 0.1036361
#>  [6,] 0.2091329 2.3556341 1.6269640
#>  [7,] 1.8569249 2.2756860 2.4874492
#>  [8,] 1.6521221 0.1652535 0.8968229
#>  [9,] 0.5032529 2.6018033 1.1345548
#> [10,] 0.8235743 0.1655401 0.6197870

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1446536 1.58098096 0.29247687
#>  [2,] 1.8865301 1.79872808 0.54388629
#>  [3,] 0.8607428 0.17465945 0.76207612
#>  [4,] 4.6733735 0.39014243 0.63950154
#>  [5,] 0.5291720 0.99145147 1.53598879
#>  [6,] 0.7632805 0.76328045 0.76328045
#>  [7,] 0.5371069 0.54876310 0.40888257
#>  [8,] 1.6162689 6.35836701 0.63829842
#>  [9,] 3.9773952 0.02019013 0.09124014
#> [10,] 0.1741535 0.90093585 2.23363184
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.15280919 2.15280919 2.15280919
#>  [2,] 0.37989983 0.37989983 0.37989983
#>  [3,] 0.73263612 0.73263612 0.73263612
#>  [4,] 0.79435647 0.79435647 0.79435647
#>  [5,] 0.17341019 0.17341019 0.17341019
#>  [6,] 0.82419501 0.82419501 0.82419501
#>  [7,] 0.08767735 0.08767735 0.08767735
#>  [8,] 0.69383028 0.69383028 0.69383028
#>  [9,] 1.62902645 1.62902645 1.62902645
#> [10,] 2.06366190 2.06366190 2.06366190
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 3.32436915 0.87292192 1.00048739
#>  [2,] 0.02948483 0.15884258 0.09605503
#>  [3,] 0.49619398 0.33641346 0.78881467
#>  [4,] 0.94051457 0.05467518 0.16333059
#>  [5,] 1.46166883 0.47854517 0.13300204
#>  [6,] 0.69326034 0.67881435 0.23208176
#>  [7,] 1.37365280 0.09547427 0.05973613
#>  [8,] 0.18181398 3.13982746 0.07782021
#>  [9,] 0.35885665 0.32456207 0.91650585
#> [10,] 0.96497862 3.25749303 1.12622716

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3590295 2.6694206 0.7443875
#>  [2,] 0.9812713 1.8080835 0.8429796
#>  [3,] 1.5864403 6.5661479 0.4475497
#>  [4,] 3.4180005 4.8253618 1.5449598
#>  [5,] 0.9063218 4.8333031 0.1277463
#>  [6,] 3.3522020 0.8728207 2.1049387
#>  [7,] 1.1111589 1.1111589 0.1647993
#>  [8,] 0.4579177 2.3449175 1.1082096
#>  [9,] 0.7317401 0.7317401 0.7317401
#> [10,] 2.7679794 2.5491855 1.3425568
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0241753 2.0241753 2.0241753
#>  [2,] 0.4906743 0.4906743 0.4906743
#>  [3,] 0.7310412 0.7310412 0.7310412
#>  [4,] 0.1927127 0.1927127 0.1927127
#>  [5,] 4.9197688 4.9197688 4.9197688
#>  [6,] 0.6531167 0.6531167 0.6531167
#>  [7,] 2.1672972 2.1672972 2.1672972
#>  [8,] 0.0602481 0.0602481 0.0602481
#>  [9,] 0.5257276 0.5257276 0.5257276
#> [10,] 0.8745811 0.8745811 0.8745811
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.05067384 0.08368415 0.9223864
#>  [2,] 0.43633769 0.46790022 1.6899636
#>  [3,] 1.66785531 0.47091813 1.4859147
#>  [4,] 0.06423428 0.84802224 2.8266592
#>  [5,] 0.39584580 1.18055206 2.1372522
#>  [6,] 1.09236438 0.09499302 0.2899166
#>  [7,] 0.23435675 0.40396651 0.8927084
#>  [8,] 0.34761669 0.76425008 1.4855215
#>  [9,] 0.26185804 0.69744653 0.1395986
#> [10,] 0.36310395 0.49221863 0.5792711

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.90638149 0.3627710 1.0191138
#>  [2,] 1.24110030 1.2411003 1.2411003
#>  [3,] 0.17672177 0.1959957 0.1959957
#>  [4,] 0.18927257 0.2919159 4.9843878
#>  [5,] 0.05922912 0.1227003 0.0935047
#>  [6,] 1.46523905 1.4652391 0.9730036
#>  [7,] 3.30586550 0.6629923 0.9720305
#>  [8,] 0.85241589 0.3780912 0.3558658
#>  [9,] 0.45605856 0.2145508 2.0921570
#> [10,] 0.70970460 0.0977441 0.1922545
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.54215008 0.54215008 0.54215008
#>  [2,] 2.47253892 2.47253892 2.47253892
#>  [3,] 2.74422011 2.74422011 2.74422011
#>  [4,] 1.97361873 1.97361873 1.97361873
#>  [5,] 0.01019553 0.01019553 0.01019553
#>  [6,] 0.30655622 0.30655622 0.30655622
#>  [7,] 1.59831422 1.59831422 1.59831422
#>  [8,] 0.71141138 0.71141138 0.71141138
#>  [9,] 0.39762483 0.39762483 0.39762483
#> [10,] 1.33134358 1.33134358 1.33134358
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.81511691 0.2007849 3.2492541
#>  [2,] 1.31308399 1.1952741 1.2803728
#>  [3,] 0.48642963 0.5216944 0.3809960
#>  [4,] 2.97996582 0.5527680 0.3560452
#>  [5,] 3.44091225 1.0661489 1.2455631
#>  [6,] 0.94770676 1.9870292 0.4073027
#>  [7,] 0.04925358 2.8889026 0.6660554
#>  [8,] 0.77270554 0.7749331 2.5755074
#>  [9,] 1.11721876 0.6058800 2.7587010
#> [10,] 0.38184559 0.3180681 0.4518020

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.81802629 0.8180263 0.8180263
#>  [2,] 0.22018958 0.7075575 0.6813766
#>  [3,] 1.86069114 2.1889502 1.4176571
#>  [4,] 0.55555033 0.3627532 1.5119628
#>  [5,] 0.25894132 1.9358001 0.4977886
#>  [6,] 3.75863343 0.6226887 1.2243072
#>  [7,] 1.74122834 2.3574112 3.2053142
#>  [8,] 0.45618044 0.4561804 0.4561804
#>  [9,] 2.69503460 2.6950346 2.6950346
#> [10,] 0.01754125 0.7652668 0.8177646
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05431771 0.05431771 0.05431771
#>  [2,] 0.36760328 0.36760328 0.36760328
#>  [3,] 0.21238436 0.21238436 0.21238436
#>  [4,] 0.25246763 0.25246763 0.25246763
#>  [5,] 0.14865052 0.14865052 0.14865052
#>  [6,] 0.34067535 0.34067535 0.34067535
#>  [7,] 1.44004598 1.44004598 1.44004598
#>  [8,] 0.34573400 0.34573400 0.34573400
#>  [9,] 1.68216072 1.68216072 1.68216072
#> [10,] 0.54561339 0.54561339 0.54561339
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.61262076 0.20723154 2.0787371
#>  [2,] 1.32597103 0.23006323 1.2303914
#>  [3,] 1.16916116 0.97564037 1.0048995
#>  [4,] 1.55336107 2.76408584 3.6492886
#>  [5,] 0.06402088 0.48146318 1.4890084
#>  [6,] 0.48151439 0.80982222 0.9220771
#>  [7,] 0.53382073 0.92864842 0.2414535
#>  [8,] 0.16899827 0.05798430 1.3274247
#>  [9,] 0.63155792 0.16215095 4.2157811
#> [10,] 0.92900676 0.06192594 1.7620572

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.29118913 2.15654772 2.08827081
#>  [2,] 1.17435900 1.17435900 1.17435900
#>  [3,] 0.02388307 0.09698151 0.47531375
#>  [4,] 0.44058278 3.61141733 1.21220077
#>  [5,] 0.03309943 0.03309943 0.03309943
#>  [6,] 0.21529851 0.66180129 0.94986530
#>  [7,] 0.35092913 0.98006945 0.98006945
#>  [8,] 1.36180812 4.26338469 3.66413767
#>  [9,] 2.35705128 2.35705128 1.76692172
#> [10,] 0.30303136 0.41473950 2.53274983
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.52652518 0.52652518 0.52652518
#>  [2,] 0.93743477 0.93743477 0.93743477
#>  [3,] 0.73738709 0.73738709 0.73738709
#>  [4,] 0.28737891 0.28737891 0.28737891
#>  [5,] 0.84021755 0.84021755 0.84021755
#>  [6,] 0.53749404 0.53749404 0.53749404
#>  [7,] 0.28669822 0.28669822 0.28669822
#>  [8,] 0.09457105 0.09457105 0.09457105
#>  [9,] 0.26420892 0.26420892 0.26420892
#> [10,] 0.07797809 0.07797809 0.07797809
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>              [,1]      [,2]       [,3]
#>  [1,] 0.001319073 0.5038225 2.34713385
#>  [2,] 0.568003995 0.5451766 0.97336269
#>  [3,] 1.134376562 1.5170983 0.37735938
#>  [4,] 0.165501632 0.6080259 1.41831201
#>  [5,] 1.029993361 0.1420929 1.80420427
#>  [6,] 2.211195684 3.1320808 0.01755667
#>  [7,] 2.108729908 0.2811949 1.84182596
#>  [8,] 1.240810429 1.1189443 0.72974392
#>  [9,] 0.636457404 1.2038529 0.03353786
#> [10,] 0.270570989 0.7396608 0.06887796

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]       [,2]        [,3]
#>  [1,] 3.2797528 3.11893963 0.346935779
#>  [2,] 1.7912833 0.72316377 0.104704461
#>  [3,] 0.6618018 0.45565901 0.661801838
#>  [4,] 0.4282525 5.28858490 0.795459275
#>  [5,] 2.1977689 2.88617274 0.001256398
#>  [6,] 2.8348291 1.48324322 1.943189663
#>  [7,] 1.3860430 1.86172797 0.588838126
#>  [8,] 0.5671024 1.96807957 0.250457982
#>  [9,] 1.7199922 2.64268573 0.107429550
#> [10,] 5.0830482 0.04882761 2.955389305
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5481960 0.5481960 0.5481960
#>  [2,] 2.1296633 2.1296633 2.1296633
#>  [3,] 0.1006494 0.1006494 0.1006494
#>  [4,] 1.9937102 1.9937102 1.9937102
#>  [5,] 0.1459904 0.1459904 0.1459904
#>  [6,] 2.3870350 2.3870350 2.3870350
#>  [7,] 0.7340863 0.7340863 0.7340863
#>  [8,] 0.5433905 0.5433905 0.5433905
#>  [9,] 0.9157642 0.9157642 0.9157642
#> [10,] 2.4955412 2.4955412 2.4955412
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9085804 0.1024112 0.2605378
#>  [2,] 0.5698397 0.2208433 0.4839629
#>  [3,] 1.1353179 0.8161932 3.7402803
#>  [4,] 1.0701937 1.5609121 0.6924920
#>  [5,] 1.0871788 0.7043499 1.3745993
#>  [6,] 0.7340066 2.7085186 3.7437313
#>  [7,] 0.4686293 1.1066030 0.2414390
#>  [8,] 1.1433467 1.7263443 0.2521602
#>  [9,] 0.0220860 1.2530717 1.2027991
#> [10,] 0.6916060 2.2381236 1.2298996

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01224437 0.97102066 0.81012685
#>  [2,] 1.49544244 0.97728007 5.42403737
#>  [3,] 0.04889951 0.08237861 0.06817342
#>  [4,] 0.17786323 0.34119517 3.27900277
#>  [5,] 0.22233523 0.22233523 0.22233523
#>  [6,] 1.46641850 1.99761041 1.99761041
#>  [7,] 0.51633095 5.29260335 4.53108100
#>  [8,] 0.24427202 0.24427202 0.24427202
#>  [9,] 1.61949893 0.58667622 0.06563355
#> [10,] 0.74105337 0.19609885 0.74105337
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9598377 0.9598377 0.9598377
#>  [2,] 0.5189195 0.5189195 0.5189195
#>  [3,] 3.1977293 3.1977293 3.1977293
#>  [4,] 4.1645609 4.1645609 4.1645609
#>  [5,] 0.6460368 0.6460368 0.6460368
#>  [6,] 0.9275970 0.9275970 0.9275970
#>  [7,] 1.6112404 1.6112404 1.6112404
#>  [8,] 2.5678758 2.5678758 2.5678758
#>  [9,] 1.9498940 1.9498940 1.9498940
#> [10,] 0.6088800 0.6088800 0.6088800
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1001604 0.3874936 0.09891081
#>  [2,] 0.4154887 0.4744215 0.83537517
#>  [3,] 1.3395404 0.7143857 5.08857310
#>  [4,] 0.1661813 2.0980010 1.34336109
#>  [5,] 0.1140039 0.7270736 2.07647541
#>  [6,] 4.5479888 1.0843472 2.12946095
#>  [7,] 0.1540127 0.9219954 1.01767121
#>  [8,] 3.6682848 1.4267537 2.99349736
#>  [9,] 0.1396630 0.6462653 0.24137687
#> [10,] 0.2947218 0.1128476 0.49330091

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.508169000 0.281988165 0.281988165
#>  [2,] 0.116610693 0.212139669 0.156802382
#>  [3,] 0.005709524 0.005709524 0.005709524
#>  [4,] 0.904110169 0.921868337 0.091337523
#>  [5,] 0.423335289 0.828374327 0.769172253
#>  [6,] 0.639215616 0.346406221 1.048567417
#>  [7,] 0.355483771 1.670358783 0.400001836
#>  [8,] 0.017729424 0.017729424 0.017729424
#>  [9,] 0.136762123 0.136762123 0.136762123
#> [10,] 0.304195514 0.845578437 0.460378215
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20695232 0.94273747 0.11489765
#>  [2,] 0.09821809 0.30367738 0.30367738
#>  [3,] 0.60600769 0.60600769 0.60600769
#>  [4,] 0.45473923 0.12540772 0.34071426
#>  [5,] 0.74338441 0.45383377 0.21109828
#>  [6,] 0.55022536 0.18160083 0.18160083
#>  [7,] 1.34560301 1.34560301 0.08097446
#>  [8,] 0.52186127 0.08904843 0.89479237
#>  [9,] 0.45330478 0.45330478 0.45330478
#> [10,] 0.29457887 0.29457887 0.29457887
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.03434048 0.38092792 0.009822439
#>  [2,] 0.39286813 0.39286813 0.162695612
#>  [3,] 0.60601843 0.60601843 1.137764267
#>  [4,] 0.48180754 0.71278547 0.051748606
#>  [5,] 0.01483101 0.01483101 0.014831011
#>  [6,] 0.07711627 1.16735155 1.255295140
#>  [7,] 0.86281799 0.86281799 0.862817985
#>  [8,] 1.04384533 0.55306278 1.043845330
#>  [9,] 0.53113944 0.08557506 2.769286248
#> [10,] 0.10633053 0.10633053 0.106330528
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31078668 0.31078668 0.05161712
#>  [2,] 1.36375992 1.31319460 0.35255898
#>  [3,] 1.39016521 0.04243724 0.47926795
#>  [4,] 0.13773334 0.05268902 0.05268902
#>  [5,] 0.53448734 0.53448734 0.53448734
#>  [6,] 0.35439570 1.06936039 0.24971903
#>  [7,] 0.14141984 0.24645555 0.90408502
#>  [8,] 0.20990654 0.20990654 1.03862203
#>  [9,] 0.01120744 0.10194114 1.28876184
#> [10,] 0.94102042 0.75164579 0.52350893
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7741936 0.2516269 0.6612101
#>  [2,] 1.4936410 0.5979205 1.4936410
#>  [3,] 1.6658998 0.2251033 0.2251033
#>  [4,] 0.2144950 0.9067977 0.1137433
#>  [5,] 0.8327617 1.0405504 0.7615256
#>  [6,] 0.9378977 0.6581243 0.9378977
#>  [7,] 0.2321101 1.4730937 0.8258991
#>  [8,] 0.5821449 0.5821449 0.5821449
#>  [9,] 1.0050530 1.0351808 0.2399007
#> [10,] 1.2429779 1.2429779 0.1909932

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]        [,3]
#>  [1,] 1.9181561 1.0134998 4.478756039
#>  [2,] 2.7958753 1.7698518 0.326878350
#>  [3,] 0.2097143 0.1588906 0.224808573
#>  [4,] 5.3553102 2.6867268 0.184893567
#>  [5,] 3.7024801 0.1267451 0.631783873
#>  [6,] 1.4581178 4.7211344 0.006708706
#>  [7,] 3.5002890 1.8971099 0.211127441
#>  [8,] 0.4932232 2.7682933 3.172792095
#>  [9,] 1.5314435 0.1147290 1.841093301
#> [10,] 0.2902819 1.3050956 1.443721278
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2601388 1.1939218 0.02648402
#>  [2,] 0.1360537 0.1360537 0.13605374
#>  [3,] 0.9198197 1.8018620 1.80186204
#>  [4,] 0.5860152 1.1501092 1.54839105
#>  [5,] 0.1874818 0.1874818 0.18748175
#>  [6,] 1.0665313 0.6689179 0.68147508
#>  [7,] 0.6877925 1.1320658 1.13206582
#>  [8,] 1.1447568 2.4753276 2.55844680
#>  [9,] 2.7835003 1.1626293 0.39380816
#> [10,] 2.4874486 1.3261652 0.97735156
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41961131 2.59457496 0.39850580
#>  [2,] 0.30369889 0.30369889 0.30369889
#>  [3,] 0.93365645 6.57515550 0.46403820
#>  [4,] 0.22174038 0.29025383 0.11932403
#>  [5,] 0.53819118 0.53819118 0.53819118
#>  [6,] 0.31746575 4.17523306 1.32558769
#>  [7,] 0.09547333 0.09547333 0.09547333
#>  [8,] 0.52466338 0.52466338 0.52466338
#>  [9,] 3.24408343 2.04721061 0.62329787
#> [10,] 0.39864466 0.43139512 1.20415378
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3712851 3.5366093 0.5883772
#>  [2,] 0.5749426 1.5890282 1.0770727
#>  [3,] 4.8421452 1.4484226 0.7090306
#>  [4,] 0.7253840 0.7253840 0.7253840
#>  [5,] 2.1432891 1.8694487 0.8851012
#>  [6,] 3.6537899 2.1759269 2.1371346
#>  [7,] 1.6691885 0.2427922 3.3684586
#>  [8,] 0.8341663 1.9221776 0.2868651
#>  [9,] 2.2544652 2.2438094 0.4638077
#> [10,] 0.5218979 0.7990096 0.4216461
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8217879 1.5770385 4.86160438
#>  [2,] 0.8423669 1.2697096 1.66861111
#>  [3,] 0.2857175 0.2857175 0.28571750
#>  [4,] 1.8772702 1.9183319 0.25235074
#>  [5,] 0.1868159 1.9836699 3.91679633
#>  [6,] 0.8516525 1.8974198 1.72983791
#>  [7,] 1.4073964 0.4381884 0.48695538
#>  [8,] 1.1196179 1.1196179 0.07444355
#>  [9,] 0.2050319 0.2050319 0.20503193
#> [10,] 0.1635941 0.1635941 0.05042908

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.0589718 0.51548067 1.6827270
#>  [2,] 0.6013273 0.60132728 0.6013273
#>  [3,] 0.2862996 0.28629960 0.2448509
#>  [4,] 0.6633684 1.42283832 1.4221678
#>  [5,] 1.4672189 0.14231142 0.1558452
#>  [6,] 1.6688259 0.10988264 0.1098826
#>  [7,] 0.4564844 0.79823015 0.7800297
#>  [8,] 0.3415233 0.08039988 0.8617429
#>  [9,] 1.9778156 0.45003247 0.9141415
#> [10,] 0.5493361 0.54933609 0.5493361
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>            [,1]        [,2]      [,3]
#>  [1,] 0.3676983 0.262205301 0.3676983
#>  [2,] 0.5469525 1.139473764 0.1130610
#>  [3,] 0.5321946 0.045949683 0.7731487
#>  [4,] 0.4122563 0.412256307 0.4122563
#>  [5,] 0.3381880 0.106753500 0.4464176
#>  [6,] 0.5537126 0.088119314 0.5245010
#>  [7,] 0.1756010 0.002863632 1.2195404
#>  [8,] 0.4546148 0.454614793 0.4546148
#>  [9,] 0.4178888 0.417888769 0.4178888
#> [10,] 0.8313717 0.831371682 0.8313717
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.67814046 0.660294480 0.21854126
#>  [2,] 0.65060260 0.001493859 0.02474951
#>  [3,] 0.74782180 0.104194341 0.57894130
#>  [4,] 0.82585729 0.825857294 0.82585729
#>  [5,] 0.20626162 0.206261618 0.20626162
#>  [6,] 0.95828049 0.501270571 0.50127057
#>  [7,] 0.64408525 0.338668524 0.33866852
#>  [8,] 0.09311370 0.093113703 0.09311370
#>  [9,] 0.02793715 0.027937149 1.10791165
#> [10,] 0.17639219 0.053449436 0.17639219
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>               [,1]       [,2]         [,3]
#>  [1,] 1.1248275037 1.12482750 0.6501489516
#>  [2,] 0.1146744746 0.48646260 0.4864625971
#>  [3,] 0.3673750856 0.03540797 0.3673750856
#>  [4,] 0.1578225170 0.15782252 0.1578225170
#>  [5,] 0.0973603955 0.09985556 0.4770942431
#>  [6,] 0.0001532543 0.45832474 0.0001532543
#>  [7,] 0.6105890698 0.61058907 0.1359747195
#>  [8,] 0.6429362249 0.64293622 0.6429362249
#>  [9,] 0.1150622956 0.11506230 0.1150622956
#> [10,] 0.1316535787 0.18451909 0.1845190922
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9921026 0.99210265 0.2519799
#>  [2,] 0.8325405 1.50979275 0.3505045
#>  [3,] 0.1039422 1.85930660 0.1329739
#>  [4,] 0.5849040 0.58490395 0.5849040
#>  [5,] 0.1744805 0.17448047 0.1744805
#>  [6,] 0.9149631 0.42417230 0.1631515
#>  [7,] 0.4870210 0.03323782 1.1551531
#>  [8,] 0.1438318 0.14383177 0.1438318
#>  [9,] 0.6725647 0.84303026 0.3048695
#> [10,] 0.3258486 0.32584864 0.3258486

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.04786151 1.0997904 0.26275759
#>  [2,] 0.23115883 0.1649433 0.23115883
#>  [3,] 0.13930309 0.1393031 0.24791660
#>  [4,] 1.46047098 1.0437853 0.62639050
#>  [5,] 0.53127132 0.6664556 0.05656505
#>  [6,] 0.23277769 0.4074479 0.24834699
#>  [7,] 0.93131377 0.1019179 0.19204437
#>  [8,] 0.12560632 0.6011064 0.12661300
#>  [9,] 0.17219269 0.1721927 0.17219269
#> [10,] 0.23074627 0.2307463 0.23074627
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>            [,1]      [,2]        [,3]
#>  [1,] 0.7557727 1.0178314 1.149449959
#>  [2,] 0.1872591 0.4386920 0.005293087
#>  [3,] 0.5008082 1.3871254 0.500808212
#>  [4,] 0.2036672 0.2036672 0.203667232
#>  [5,] 0.9938942 0.2656178 0.954069382
#>  [6,] 1.3592628 1.7249361 0.108594815
#>  [7,] 0.1994702 0.1994702 0.326278002
#>  [8,] 0.3824684 0.3824684 0.382468396
#>  [9,] 0.2099279 0.6969826 0.275818216
#> [10,] 0.2079643 0.2079643 0.207964343
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02611167 0.11991088 0.40416784
#>  [2,] 0.02125511 0.30650178 0.02125511
#>  [3,] 0.58460607 0.58460607 0.58460607
#>  [4,] 0.53120103 0.07304791 0.31258043
#>  [5,] 0.06444191 0.06444191 0.06444191
#>  [6,] 0.73232372 1.48617895 0.93684899
#>  [7,] 0.14644921 0.42653301 0.02630822
#>  [8,] 0.51203158 0.29790114 0.70158212
#>  [9,] 0.88160584 0.39842548 0.16466691
#> [10,] 1.27672734 0.32968443 1.27672734
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.09726074 0.8673518 1.0743161
#>  [2,] 0.03631745 0.2253725 0.2962268
#>  [3,] 1.13509202 0.2056478 0.7257466
#>  [4,] 0.31678315 0.3955322 0.9066752
#>  [5,] 0.10403063 0.4243423 0.2524120
#>  [6,] 0.19657265 0.2696762 0.2234596
#>  [7,] 0.19074676 0.5831205 0.3511645
#>  [8,] 0.24299843 0.6275550 2.4745769
#>  [9,] 0.18253742 0.2473036 0.2473036
#> [10,] 0.21505376 0.2150538 0.2150538

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02875897 0.71020788 0.71020788
#>  [2,] 0.72285229 0.01455587 0.72285229
#>  [3,] 0.16473444 0.16473444 0.08461972
#>  [4,] 0.48323684 0.08108465 0.48323684
#>  [5,] 0.55075685 0.16849298 0.17108992
#>  [6,] 0.66732893 0.29839905 0.29839905
#>  [7,] 0.61533085 0.61320450 1.15819928
#>  [8,] 0.46525712 0.14433582 0.38949051
#>  [9,] 0.19863217 0.11516683 0.14700105
#> [10,] 0.36952895 0.50210229 0.08332656
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5813706 0.24232949 0.26489512
#>  [2,] 0.6172100 0.72305508 0.72305508
#>  [3,] 0.1725178 0.06153705 0.35475516
#>  [4,] 0.3979405 0.39794052 0.22698631
#>  [5,] 0.1386810 0.09328491 0.19993519
#>  [6,] 0.6865467 0.03534029 0.09319508
#>  [7,] 1.0899700 1.41524584 0.38751082
#>  [8,] 0.3184801 0.31848006 0.42360823
#>  [9,] 0.4296410 0.42964102 0.42964102
#> [10,] 0.5879293 0.04860395 0.01243946
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.086615674 0.06378580 0.485922540
#>  [2,] 0.993817716 1.11141758 0.058272443
#>  [3,] 0.051587184 0.07911996 0.051587184
#>  [4,] 0.486926385 0.24011764 0.256371607
#>  [5,] 0.008713193 0.06678452 0.008713193
#>  [6,] 0.080066809 0.08006681 0.080066809
#>  [7,] 0.017118803 0.01711880 0.017118803
#>  [8,] 0.059228034 0.37127235 0.206476668
#>  [9,] 0.015197563 0.46417782 0.015197563
#> [10,] 0.035123953 1.09656626 0.315893844
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.08426941 0.153162444 0.08426941
#>  [2,] 0.46088662 0.125426206 0.45980262
#>  [3,] 0.98850728 0.512251898 0.51225190
#>  [4,] 0.14995296 0.001252592 0.27443403
#>  [5,] 0.66902545 0.523380265 0.21527956
#>  [6,] 0.14722431 0.063862918 0.32323547
#>  [7,] 0.28356593 0.283565927 0.97290397
#>  [8,] 0.30051577 0.300515771 0.30051577
#>  [9,] 0.17583953 0.143676893 0.17583953
#> [10,] 0.19999895 0.158874492 0.38977975

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.22850501 0.22907444 0.3809877
#>  [2,] 0.27237964 0.52737516 0.9661834
#>  [3,] 0.14786496 0.14786496 0.1478650
#>  [4,] 0.16306193 0.16306193 0.1630619
#>  [5,] 0.25958835 0.86221935 0.1708417
#>  [6,] 0.14457587 0.28429757 0.0763176
#>  [7,] 1.82077244 0.13434818 0.1343482
#>  [8,] 0.03771157 0.03771157 0.5198529
#>  [9,] 0.33920657 0.41118110 0.4511800
#> [10,] 0.44628231 0.57307604 0.4462823
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]         [,2]      [,3]
#>  [1,] 0.48757998 0.4875799798 0.4875800
#>  [2,] 0.10763089 0.1076308945 0.1076309
#>  [3,] 0.12205290 0.0747201417 0.1220529
#>  [4,] 0.27462062 0.1568459568 0.1568460
#>  [5,] 0.26087527 0.1118509565 0.5057098
#>  [6,] 0.69945743 0.0006417445 0.4955545
#>  [7,] 0.23081609 0.2308160879 0.4392869
#>  [8,] 0.19805999 1.2054860831 0.1543122
#>  [9,] 0.01487427 0.8326094774 0.5354179
#> [10,] 0.63884931 0.3918688420 0.1875566
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.173033621 1.14933675 0.43860985
#>  [2,] 0.826332929 0.91670570 0.82278619
#>  [3,] 0.197067968 0.19706797 0.19706797
#>  [4,] 0.321005611 0.49204254 0.32100561
#>  [5,] 0.617893293 0.10730035 0.10730035
#>  [6,] 0.006891133 0.03509934 0.09302418
#>  [7,] 0.241146936 1.01002210 0.24928631
#>  [8,] 0.009668160 0.22703042 0.27237562
#>  [9,] 0.959502073 1.45569612 0.34030802
#> [10,] 0.006192007 0.20690719 0.20690719
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.441844896 0.058029702 0.904440164
#>  [2,] 0.088563502 0.008628423 0.477832877
#>  [3,] 0.448624996 1.176581009 0.530973280
#>  [4,] 0.003818617 0.003818617 0.003818617
#>  [5,] 0.088740949 0.088740949 0.088740949
#>  [6,] 0.043380226 0.027677293 0.324315224
#>  [7,] 0.105080774 0.140510910 0.140510910
#>  [8,] 0.268301102 0.187963553 0.070542436
#>  [9,] 0.049378208 0.049378208 0.049378208
#> [10,] 0.108349958 0.085636915 0.108349958