Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07562247 0.06880376 1.22226369
#>  [2,] 0.10660756 1.75135984 3.04485709
#>  [3,] 1.32282081 1.32282081 1.14488060
#>  [4,] 1.39625624 0.61297481 1.84353269
#>  [5,] 0.96858910 3.54187678 1.41281246
#>  [6,] 2.20572474 1.08461559 0.31077798
#>  [7,] 0.10706817 4.21558553 1.41807250
#>  [8,] 5.31680893 2.10595361 0.07939456
#>  [9,] 0.38962916 0.83415653 1.26932662
#> [10,] 1.65850246 1.76888748 3.78411706
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4199103 0.4199103 0.4199103
#>  [2,] 0.4381556 0.4381556 0.4381556
#>  [3,] 1.4045021 1.4045021 1.4045021
#>  [4,] 0.9186128 0.9186128 0.9186128
#>  [5,] 1.8534036 1.8534036 1.8534036
#>  [6,] 2.8533372 2.8533372 2.8533372
#>  [7,] 0.2189438 0.2189438 0.2189438
#>  [8,] 0.3811815 0.3811815 0.3811815
#>  [9,] 0.2215447 0.2215447 0.2215447
#> [10,] 0.2494086 0.2494086 0.2494086
## independence
rpextmo(10, 3, b = 1)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.358765026 0.07089188 2.88084984
#>  [2,] 3.180007947 1.27445240 0.50929147
#>  [3,] 0.001346781 2.37453202 2.94231591
#>  [4,] 0.213970551 0.68446770 0.28397372
#>  [5,] 0.774385868 1.84933316 0.72945303
#>  [6,] 0.103432404 0.38413216 0.41872929
#>  [7,] 0.053177528 1.46693442 0.05768267
#>  [8,] 0.062161516 2.12657290 0.52887491
#>  [9,] 2.046246340 0.23229636 0.41761608
#> [10,] 0.394675206 1.01466500 2.84377215

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]        [,3]
#>  [1,] 0.1213434 0.1213434 0.121343384
#>  [2,] 0.4460246 0.3455336 0.918671051
#>  [3,] 0.3613460 0.7310589 0.006498462
#>  [4,] 0.3131166 0.3131166 0.257858017
#>  [5,] 1.8496293 1.8496293 0.195170308
#>  [6,] 0.1820430 2.1499032 1.441790157
#>  [7,] 4.9882030 4.5750548 3.648136236
#>  [8,] 0.9073678 0.0751811 4.085214834
#>  [9,] 1.1868477 0.5204738 0.384737083
#> [10,] 2.3310268 0.1437260 1.284132896
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18352205 0.18352205 0.18352205
#>  [2,] 1.13118659 1.13118659 1.13118659
#>  [3,] 1.14752761 1.14752761 1.14752761
#>  [4,] 1.97810335 1.97810335 1.97810335
#>  [5,] 1.08325641 1.08325641 1.08325641
#>  [6,] 1.29649035 1.29649035 1.29649035
#>  [7,] 0.06453504 0.06453504 0.06453504
#>  [8,] 1.25209602 1.25209602 1.25209602
#>  [9,] 3.88622076 3.88622076 3.88622076
#> [10,] 1.41328818 1.41328818 1.41328818
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.09890077 2.39642405 0.28211715
#>  [2,] 2.19011474 0.15913524 4.32925596
#>  [3,] 1.12048803 0.93898218 0.73443205
#>  [4,] 0.69185122 0.37447799 0.83229418
#>  [5,] 3.11990073 1.07790117 0.60381018
#>  [6,] 0.64472328 0.14501255 2.18537140
#>  [7,] 2.54807905 1.90682764 0.46094318
#>  [8,] 0.20913812 0.44320940 0.64487025
#>  [9,] 1.72615625 1.46487084 0.09303694
#> [10,] 0.03821058 0.05473967 0.79730593

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3552909 0.09679808 3.26062823
#>  [2,] 0.9286246 0.24868352 2.06116359
#>  [3,] 1.5452385 1.08472706 1.50085256
#>  [4,] 0.1204863 0.12048628 0.12048628
#>  [5,] 1.0403062 1.04030619 1.04030619
#>  [6,] 0.6737827 1.78706669 1.01607503
#>  [7,] 0.9848223 0.95370468 0.04032741
#>  [8,] 2.0390856 0.16498011 2.03908558
#>  [9,] 2.4419384 2.44193836 0.34807127
#> [10,] 0.1088417 0.10884166 0.10884166
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.29536309 2.29536309 2.29536309
#>  [2,] 0.87784029 0.87784029 0.87784029
#>  [3,] 0.12644257 0.12644257 0.12644257
#>  [4,] 0.22885367 0.22885367 0.22885367
#>  [5,] 2.65596265 2.65596265 2.65596265
#>  [6,] 1.59951374 1.59951374 1.59951374
#>  [7,] 0.09351488 0.09351488 0.09351488
#>  [8,] 1.14435627 1.14435627 1.14435627
#>  [9,] 0.19616741 0.19616741 0.19616741
#> [10,] 0.34857681 0.34857681 0.34857681
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.91393713 0.06823392 0.1536650
#>  [2,] 0.04852064 1.12915266 0.7492475
#>  [3,] 0.72581408 0.84897568 0.1924344
#>  [4,] 0.83554794 2.06030388 0.4249772
#>  [5,] 1.38985171 1.30442287 1.9529912
#>  [6,] 0.05326895 2.00738790 0.1338819
#>  [7,] 0.08604789 1.15319862 2.0270709
#>  [8,] 0.18323404 0.13491053 1.4363885
#>  [9,] 3.64802853 0.26879743 0.2091548
#> [10,] 0.28831139 0.23376880 0.3707761

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>               [,1]         [,2]         [,3]
#>  [1,] 0.9374710867 2.4177240578 1.4611206122
#>  [2,] 0.1764311515 0.1764311515 0.1764311515
#>  [3,] 0.2880908644 0.0607567735 0.1236279131
#>  [4,] 0.6781046317 0.6781046317 0.6781046317
#>  [5,] 0.0531439500 2.5780465009 0.9678660625
#>  [6,] 0.8110440355 0.5976190511 1.0821984733
#>  [7,] 3.6336202134 0.8750993833 0.3691720979
#>  [8,] 0.0008307373 0.0008307373 0.0008307373
#>  [9,] 0.2461841683 0.3016488254 0.3016488254
#> [10,] 2.8984115006 0.3761794388 0.2127166138
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01184496 0.01184496 0.01184496
#>  [2,] 0.12294437 0.12294437 0.12294437
#>  [3,] 0.67727804 0.67727804 0.67727804
#>  [4,] 1.09983920 1.09983920 1.09983920
#>  [5,] 0.46510937 0.46510937 0.46510937
#>  [6,] 0.62905440 0.62905440 0.62905440
#>  [7,] 1.25890430 1.25890430 1.25890430
#>  [8,] 0.03095443 0.03095443 0.03095443
#>  [9,] 1.92284722 1.92284722 1.92284722
#> [10,] 1.22091191 1.22091191 1.22091191
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5902610 0.30906087 0.19036118
#>  [2,] 0.5739926 1.16525327 0.41660421
#>  [3,] 0.2103143 0.11073782 0.05918105
#>  [4,] 1.6734390 0.43733574 0.24033479
#>  [5,] 1.7352006 0.46271677 0.13584717
#>  [6,] 1.6193714 0.05474228 0.27626781
#>  [7,] 0.6324906 0.31612045 2.51454120
#>  [8,] 0.8940205 1.04818703 0.42535110
#>  [9,] 0.3002767 0.30059817 1.70463776
#> [10,] 0.2855433 0.74037448 1.25046512

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8801150 1.48265085 0.8956005
#>  [2,] 0.5858537 2.65233819 0.1238149
#>  [3,] 0.1966777 0.19667770 0.1966777
#>  [4,] 1.6168417 0.94253342 1.6168417
#>  [5,] 0.1107419 0.11074193 0.1107419
#>  [6,] 0.5724642 3.61561013 1.7499041
#>  [7,] 0.6439726 0.04873419 0.6439726
#>  [8,] 0.4238522 1.52554387 1.7188840
#>  [9,] 3.6498609 6.14623547 0.6074942
#> [10,] 4.3701289 1.31407029 0.3916065
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.509944705 0.509944705 0.509944705
#>  [2,] 3.036472089 3.036472089 3.036472089
#>  [3,] 3.379714176 3.379714176 3.379714176
#>  [4,] 0.006235987 0.006235987 0.006235987
#>  [5,] 4.751326639 4.751326639 4.751326639
#>  [6,] 0.268640551 0.268640551 0.268640551
#>  [7,] 1.087222381 1.087222381 1.087222381
#>  [8,] 0.040009255 0.040009255 0.040009255
#>  [9,] 0.175786843 0.175786843 0.175786843
#> [10,] 0.219101755 0.219101755 0.219101755
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.83551017 0.3636495 0.89960093
#>  [2,] 0.04604886 1.0185115 1.02889977
#>  [3,] 1.99204802 0.3702649 0.13611630
#>  [4,] 0.23214993 1.2068631 0.64487794
#>  [5,] 0.14414535 1.4120449 0.19889704
#>  [6,] 0.35922073 1.2919389 0.56722366
#>  [7,] 0.48793256 0.9471890 0.47440980
#>  [8,] 0.28709638 0.4303702 0.09129975
#>  [9,] 1.29097021 0.6282353 0.49622465
#> [10,] 0.16603449 0.2252467 1.56146601

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.1176859 0.9536896 3.99591605
#>  [2,] 0.5619187 0.5619187 0.56191867
#>  [3,] 1.2090103 0.4252649 2.07671119
#>  [4,] 0.2677319 0.1265008 1.64037312
#>  [5,] 0.3056758 5.0246150 0.03664916
#>  [6,] 0.1617003 2.2706090 2.59683506
#>  [7,] 1.7194323 0.3202554 0.33130184
#>  [8,] 2.8662623 0.9228857 3.69042184
#>  [9,] 0.5212464 7.5599154 1.19940875
#> [10,] 0.4043484 0.6965537 2.94186438
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6936684 0.6936684 0.6936684
#>  [2,] 0.9951738 0.9951738 0.9951738
#>  [3,] 2.4410439 2.4410439 2.4410439
#>  [4,] 1.8632709 1.8632709 1.8632709
#>  [5,] 0.9485844 0.9485844 0.9485844
#>  [6,] 3.0406769 3.0406769 3.0406769
#>  [7,] 1.9177930 1.9177930 1.9177930
#>  [8,] 0.4120170 0.4120170 0.4120170
#>  [9,] 0.3873756 0.3873756 0.3873756
#> [10,] 0.2953993 0.2953993 0.2953993
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1017161 2.86554603 0.16747773
#>  [2,] 0.6294667 0.66095222 1.24575044
#>  [3,] 0.9119728 0.35052763 0.27611596
#>  [4,] 0.7275954 0.04650877 3.05402156
#>  [5,] 0.5956668 1.77329518 0.27022095
#>  [6,] 3.3750164 2.29497566 0.06605588
#>  [7,] 0.2328394 0.67720935 0.05009549
#>  [8,] 0.7964313 0.80598373 2.87850351
#>  [9,] 1.1665264 1.43831778 0.13014972
#> [10,] 0.8317472 0.63154583 0.47899589

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]         [,2]      [,3]
#>  [1,] 0.23281873 0.0954520697 0.2328187
#>  [2,] 1.27179359 0.1311831726 4.0258432
#>  [3,] 2.57089017 0.5528333222 1.0950558
#>  [4,] 0.51070927 0.5107092671 0.3797911
#>  [5,] 2.14240457 2.3449025172 0.7608633
#>  [6,] 0.05784305 0.1610618830 0.1353696
#>  [7,] 0.99808091 0.7489526035 0.2380322
#>  [8,] 4.37805663 2.7566835041 1.1418890
#>  [9,] 0.90096428 0.0004879413 1.0011356
#> [10,] 1.89919745 1.8991974472 1.8991974
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8523131 1.8523131 1.8523131
#>  [2,] 0.4375330 0.4375330 0.4375330
#>  [3,] 0.2782487 0.2782487 0.2782487
#>  [4,] 1.8802503 1.8802503 1.8802503
#>  [5,] 1.4789918 1.4789918 1.4789918
#>  [6,] 1.1927325 1.1927325 1.1927325
#>  [7,] 1.0760445 1.0760445 1.0760445
#>  [8,] 0.3264140 0.3264140 0.3264140
#>  [9,] 0.3387460 0.3387460 0.3387460
#> [10,] 0.4269348 0.4269348 0.4269348
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84762826 0.48386508 0.37347183
#>  [2,] 0.01423934 2.23510168 3.35089091
#>  [3,] 0.38723424 0.73159827 0.85447959
#>  [4,] 0.60053927 0.23167772 2.27409246
#>  [5,] 0.48288978 0.45341096 0.07337285
#>  [6,] 0.81524763 2.83794213 1.73326672
#>  [7,] 0.22923810 0.16306378 0.01823055
#>  [8,] 1.01746941 1.14962689 0.78165743
#>  [9,] 2.47582290 0.36247009 2.82596195
#> [10,] 1.94767612 0.03781588 0.69066840

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.007781757 2.259719210 1.287944134
#>  [2,] 0.103161698 0.103161698 0.103161698
#>  [3,] 0.001327021 0.001327021 0.001327021
#>  [4,] 2.121619058 0.856475764 2.121619058
#>  [5,] 1.601128075 1.055335600 2.571213292
#>  [6,] 6.738210752 5.295217373 5.419441256
#>  [7,] 0.044536056 2.090555780 0.780145486
#>  [8,] 0.341500256 0.059061467 0.410467421
#>  [9,] 0.128961156 0.742561789 0.309017750
#> [10,] 3.009483970 0.594392717 4.610298235
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2369085 0.2369085 0.2369085
#>  [2,] 1.1954223 1.1954223 1.1954223
#>  [3,] 1.5200843 1.5200843 1.5200843
#>  [4,] 0.2717570 0.2717570 0.2717570
#>  [5,] 0.8594846 0.8594846 0.8594846
#>  [6,] 0.6014111 0.6014111 0.6014111
#>  [7,] 1.8220955 1.8220955 1.8220955
#>  [8,] 0.1831046 0.1831046 0.1831046
#>  [9,] 1.4578434 1.4578434 1.4578434
#> [10,] 1.2633398 1.2633398 1.2633398
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2019244 1.14173773 0.6489867
#>  [2,] 2.7533617 0.01593585 0.5916282
#>  [3,] 1.1726905 0.44422179 1.7550886
#>  [4,] 4.1370473 0.37105655 0.3206816
#>  [5,] 1.4321282 0.74109970 0.6332882
#>  [6,] 1.4247964 0.57382014 1.1256808
#>  [7,] 0.6911224 0.42902305 1.6765427
#>  [8,] 0.6056565 0.05776929 0.1836536
#>  [9,] 0.9278989 0.49371054 0.2003418
#> [10,] 0.0856528 0.93354941 0.6883178

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.80016705 1.89737555 1.94025687
#>  [2,] 1.40395268 2.79237994 1.34330440
#>  [3,] 0.17118197 0.17118197 0.17118197
#>  [4,] 0.07853263 1.81208536 1.81208536
#>  [5,] 0.68711379 0.39513446 0.68711379
#>  [6,] 0.06356354 1.63796508 2.06268638
#>  [7,] 1.55323057 1.47455741 1.17918991
#>  [8,] 2.03283995 1.01179584 1.01918302
#>  [9,] 0.06332144 0.06692573 0.06692573
#> [10,] 0.34036617 0.84096420 1.16538120
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.58025402 1.58025402 1.58025402
#>  [2,] 1.56078541 1.56078541 1.56078541
#>  [3,] 0.94281484 0.94281484 0.94281484
#>  [4,] 0.09222885 0.09222885 0.09222885
#>  [5,] 1.56389481 1.56389481 1.56389481
#>  [6,] 0.15732738 0.15732738 0.15732738
#>  [7,] 0.62495867 0.62495867 0.62495867
#>  [8,] 0.37420429 0.37420429 0.37420429
#>  [9,] 0.96651324 0.96651324 0.96651324
#> [10,] 0.01327591 0.01327591 0.01327591
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.09537408 1.4823219 2.02951432
#>  [2,] 1.72621982 3.9411163 0.08679425
#>  [3,] 2.23450764 0.2338334 0.66486084
#>  [4,] 0.40863182 0.4818704 0.19797324
#>  [5,] 1.31147986 0.3489868 0.16344145
#>  [6,] 2.61889662 2.9497260 0.10527731
#>  [7,] 0.45742886 0.8220220 2.26082459
#>  [8,] 0.56225612 1.2876739 0.42102707
#>  [9,] 2.09959138 4.4585984 1.13739370
#> [10,] 1.07828626 0.3084060 1.67761505

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56355724 1.56004480 1.56004480
#>  [2,] 0.02491767 0.06023968 0.08623413
#>  [3,] 0.01389943 0.01389943 0.01389943
#>  [4,] 0.37238916 0.59791089 0.75056913
#>  [5,] 0.57345442 0.04140806 4.92972363
#>  [6,] 0.63793818 0.63793818 0.54122252
#>  [7,] 0.02065167 0.48915391 0.48915391
#>  [8,] 3.99865239 1.66009088 2.89115167
#>  [9,] 0.57481240 0.06030338 3.77805643
#> [10,] 3.10256780 0.23076756 3.61966693
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.04095309 1.04095309 1.04095309
#>  [2,] 2.21751334 2.21751334 2.21751334
#>  [3,] 0.98267721 0.98267721 0.98267721
#>  [4,] 2.13975665 2.13975665 2.13975665
#>  [5,] 1.37249514 1.37249514 1.37249514
#>  [6,] 0.01017344 0.01017344 0.01017344
#>  [7,] 0.43457254 0.43457254 0.43457254
#>  [8,] 0.08188548 0.08188548 0.08188548
#>  [9,] 0.03631872 0.03631872 0.03631872
#> [10,] 2.29075878 2.29075878 2.29075878
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.14133117 0.90979652 0.2571750
#>  [2,] 1.70798876 1.88423357 1.9175483
#>  [3,] 1.65168526 0.54796183 1.0513979
#>  [4,] 0.36174792 1.01879376 1.1476328
#>  [5,] 0.61905711 6.54649492 1.0756170
#>  [6,] 0.96283374 1.51507624 0.8444391
#>  [7,] 0.30517441 0.42620607 0.3537739
#>  [8,] 1.06330489 0.03888271 0.1019705
#>  [9,] 0.09702547 0.40565388 1.8621001
#> [10,] 0.33193918 0.06969601 0.5113883

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0258791 0.2860228 1.0849743
#>  [2,] 1.7461256 2.0056480 0.5273080
#>  [3,] 0.6795299 0.5033819 0.1628025
#>  [4,] 0.8082067 0.9923696 0.3850696
#>  [5,] 0.9609976 0.1611585 0.9085927
#>  [6,] 0.3680631 0.4209818 1.3334006
#>  [7,] 0.1538672 0.1538672 0.6893479
#>  [8,] 0.2421363 0.5693445 0.5693445
#>  [9,] 0.0039147 0.0039147 0.1911438
#> [10,] 0.2926667 0.2407911 0.3519837
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02622465 0.04711277 1.29778288
#>  [2,] 1.88222809 1.88222809 1.05298679
#>  [3,] 0.03372239 0.03372239 2.01440852
#>  [4,] 0.01498025 0.01498025 0.01498025
#>  [5,] 0.18119785 0.18119785 0.11680300
#>  [6,] 0.55589439 1.27006019 0.31449370
#>  [7,] 1.27050471 1.20310081 0.80743218
#>  [8,] 0.82434447 0.34549713 0.00449621
#>  [9,] 0.24586394 0.24586394 0.24586394
#> [10,] 0.15399488 0.34639348 0.34639348
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.339709569 0.081900543 0.339709569
#>  [2,] 0.195723318 0.002768953 1.158997406
#>  [3,] 1.073290540 2.813278259 2.323698239
#>  [4,] 4.040326446 1.590673601 1.175929215
#>  [5,] 0.005018841 0.005018841 0.005018841
#>  [6,] 0.291010512 0.284786888 0.284786888
#>  [7,] 1.019683415 0.591568878 0.110364405
#>  [8,] 0.002148587 0.002148587 0.002148587
#>  [9,] 0.988396188 0.600094213 0.328596779
#> [10,] 2.259884062 0.744415715 1.330923649
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2140346 0.84718955 0.38921754
#>  [2,] 0.2388158 0.23881576 0.23881576
#>  [3,] 0.1639440 0.30561732 0.08968915
#>  [4,] 1.4816475 0.87296358 1.21033521
#>  [5,] 2.9523416 3.72488001 2.96515452
#>  [6,] 0.4024117 0.36427012 0.40241167
#>  [7,] 0.6036406 0.06401352 1.36211141
#>  [8,] 1.3637461 1.36374606 1.36374606
#>  [9,] 0.4911401 0.13299640 0.38413018
#> [10,] 0.7562150 0.78183382 0.78183382
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.16187475 0.04927575 1.23546765
#>  [2,] 0.19393845 0.51688839 0.84454394
#>  [3,] 0.01212393 0.07965198 0.01095827
#>  [4,] 0.09926942 0.04359498 0.51837581
#>  [5,] 0.08540312 0.08540312 0.02347348
#>  [6,] 0.15216412 0.30403555 0.36446840
#>  [7,] 0.07952830 0.04177186 0.07952830
#>  [8,] 0.40410011 0.18910429 0.01086867
#>  [9,] 0.18341640 0.44985336 0.72814090
#> [10,] 2.05875863 3.02551511 4.15351805

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.52615091 2.45371741 0.6555683
#>  [2,] 2.12715779 2.22141423 2.2214142
#>  [3,] 0.30919568 0.70539121 1.7172789
#>  [4,] 1.97082272 0.10913266 0.5231580
#>  [5,] 0.49379336 3.08885404 0.1876332
#>  [6,] 0.07878607 1.43502405 0.2023477
#>  [7,] 0.23179373 0.18593211 0.4065931
#>  [8,] 0.16914948 0.26978520 1.9036748
#>  [9,] 0.14624924 0.08927672 0.3815941
#> [10,] 0.35309430 0.40873784 1.4838190
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.4896508 0.66339459 1.8690932
#>  [2,] 0.8732261 0.88148393 2.3488058
#>  [3,] 0.9120720 0.25549098 1.0063679
#>  [4,] 2.8524391 0.04657719 6.3347080
#>  [5,] 0.7911881 0.07980905 0.6144589
#>  [6,] 1.8111201 0.62752356 2.6737087
#>  [7,] 0.2928786 0.15042886 0.3146537
#>  [8,] 1.7356787 0.62701598 1.8312679
#>  [9,] 0.1936172 0.13961665 1.4407581
#> [10,] 1.0515154 2.51698691 2.5169869
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 3.5998801 0.02176207 2.0701407
#>  [2,] 1.3244032 1.59746425 1.1689702
#>  [3,] 1.0756126 0.73111321 0.9058157
#>  [4,] 6.2295022 5.53587815 2.6262291
#>  [5,] 1.4706587 0.89722801 3.3586418
#>  [6,] 0.4419860 3.89071989 0.7631415
#>  [7,] 2.0764238 2.02880582 2.0764238
#>  [8,] 3.7765350 0.60992848 0.6088463
#>  [9,] 1.3509022 2.13369057 2.5001484
#> [10,] 0.2227506 0.01973849 0.2227506
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4839835 0.6415571 0.06754591
#>  [2,] 1.4302099 0.5695334 0.96213882
#>  [3,] 0.6279702 2.3293783 2.66235207
#>  [4,] 1.6016578 3.4704687 0.16530996
#>  [5,] 2.7292738 0.7141589 1.79607853
#>  [6,] 0.3364485 0.2569514 0.50278149
#>  [7,] 0.5065979 0.6488220 0.27402747
#>  [8,] 1.9156751 0.2494413 0.40576152
#>  [9,] 0.7025832 1.3443302 0.22031726
#> [10,] 0.2268691 0.2268691 0.22686906
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>              [,1]      [,2]       [,3]
#>  [1,] 1.519775657 1.4171884 2.40793502
#>  [2,] 0.183032433 1.6216015 2.92936079
#>  [3,] 0.001309354 1.4047137 0.62767371
#>  [4,] 0.260375284 0.2603753 0.26037528
#>  [5,] 1.938484004 3.2283558 0.67263065
#>  [6,] 0.264297165 2.8043618 0.06996686
#>  [7,] 0.616399034 2.4614978 0.97899901
#>  [8,] 0.982672704 0.9524388 0.18872385
#>  [9,] 1.436321643 0.1836978 0.78514096
#> [10,] 4.114106007 1.5994294 1.05912629

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]         [,2]       [,3]
#>  [1,] 0.05705679 0.0570567899 0.05705679
#>  [2,] 0.43023200 0.4302319968 0.43023200
#>  [3,] 0.11458490 0.1145849014 0.11458490
#>  [4,] 0.41301698 0.0088802711 0.41301698
#>  [5,] 0.23465720 0.0003975111 0.23465720
#>  [6,] 0.74739363 0.1734981339 0.74739363
#>  [7,] 0.09872117 0.0987211729 0.09872117
#>  [8,] 1.11004950 1.1100494972 0.29151388
#>  [9,] 0.50968101 0.5096810091 0.42985720
#> [10,] 0.41627226 0.3381236310 0.33812363
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>              [,1]        [,2]       [,3]
#>  [1,] 0.167032310 0.167032310 0.09314744
#>  [2,] 1.312798781 0.971686903 1.31279878
#>  [3,] 0.045335219 0.045335219 0.04533522
#>  [4,] 0.482470764 0.532248315 0.12912171
#>  [5,] 0.057526055 0.057526055 0.05752606
#>  [6,] 0.003067892 0.003067892 0.38560366
#>  [7,] 0.361268054 2.076229225 2.07622923
#>  [8,] 1.984048070 0.156158043 0.31353778
#>  [9,] 0.214470210 0.148155876 0.21447021
#> [10,] 0.566846591 0.167971448 0.16797145
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0143194 1.0143194 1.0143194
#>  [2,] 0.3979795 0.3979795 0.2759710
#>  [3,] 0.6388929 0.6388929 0.6388929
#>  [4,] 0.1216681 0.1216681 0.1216681
#>  [5,] 0.2664964 0.2957948 0.5169714
#>  [6,] 0.8321663 0.8321663 0.8321663
#>  [7,] 0.1160106 1.0556313 1.0556313
#>  [8,] 0.9319307 0.3275923 0.7178705
#>  [9,] 1.5621685 1.8091255 0.2075995
#> [10,] 1.0069185 0.5224622 1.0325803
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3417476 0.3417476 0.3417476
#>  [2,] 0.6199694 0.1249677 0.6291275
#>  [3,] 0.2430541 0.5451411 0.5451411
#>  [4,] 0.8706966 0.8706966 0.8706966
#>  [5,] 0.2154021 0.2154021 0.2154021
#>  [6,] 0.3324460 1.5974084 1.5974084
#>  [7,] 0.4918353 1.3993408 0.6482570
#>  [8,] 0.1406323 0.8136597 0.8136597
#>  [9,] 0.1807088 0.1807088 0.1807088
#> [10,] 0.0352604 0.0352604 0.0352604
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36345587 0.36345587 0.84191097
#>  [2,] 0.73000539 1.31802830 0.73000539
#>  [3,] 0.04263234 0.86617403 0.66243660
#>  [4,] 0.65708159 0.60286852 0.65708159
#>  [5,] 0.03526797 0.03526797 0.03526797
#>  [6,] 0.60964296 0.86627057 0.86627057
#>  [7,] 0.13270772 0.13270772 0.13270772
#>  [8,] 0.22623738 0.42013071 0.88128758
#>  [9,] 0.01790410 0.01790410 0.01790410
#> [10,] 0.26524726 0.26524726 0.26524726

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3427035 0.34270353 0.34270353
#>  [2,] 0.9937365 0.15091128 0.31989400
#>  [3,] 0.1013937 0.53449088 0.53449088
#>  [4,] 0.6696676 0.95430034 0.09537795
#>  [5,] 0.1047614 0.10476137 0.10476137
#>  [6,] 0.5999998 0.05771506 0.02190556
#>  [7,] 0.7082304 0.40394120 0.34231412
#>  [8,] 0.1397982 0.10055583 0.13979825
#>  [9,] 0.0504176 0.29248073 0.29248073
#> [10,] 0.1893577 0.18935769 0.03684151
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05990234 0.16694741 0.03539711
#>  [2,] 0.56212818 1.28025474 0.61310351
#>  [3,] 0.03865327 0.05643609 0.25067333
#>  [4,] 0.21918474 0.70089080 0.69858520
#>  [5,] 0.32476716 0.45375807 0.45375807
#>  [6,] 0.52678043 0.04455571 0.52678043
#>  [7,] 0.31447720 0.31447720 0.31447720
#>  [8,] 0.41044879 0.41044879 0.41044879
#>  [9,] 0.17699159 0.17699159 0.17699159
#> [10,] 0.60221616 0.22171950 0.60221616
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29411626 0.04679832 0.04679832
#>  [2,] 0.67819693 0.61201336 0.67819693
#>  [3,] 0.07519922 0.07519922 0.07519922
#>  [4,] 0.28011349 0.55291517 0.22729106
#>  [5,] 0.12112000 0.23647721 1.22732295
#>  [6,] 0.19058561 0.19058561 0.69002611
#>  [7,] 0.40868550 0.40868550 0.16001641
#>  [8,] 0.19193056 0.87373798 0.87373798
#>  [9,] 0.42996588 0.42996588 0.30650545
#> [10,] 0.71934180 0.13661685 0.13661685
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 0.875569695 0.87556969 0.605375956
#>  [2,] 0.007936135 0.91085035 0.007936135
#>  [3,] 0.792622627 0.33627986 0.459994973
#>  [4,] 0.892013947 0.80291252 0.105198379
#>  [5,] 0.636427620 0.19108433 0.191084326
#>  [6,] 0.595455097 0.73173949 0.276123441
#>  [7,] 0.074607565 0.07460756 0.074607565
#>  [8,] 0.331494420 0.12595156 0.246971251
#>  [9,] 0.070329294 0.07032929 0.070329294
#> [10,] 0.268009300 0.16458127 0.268009300

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03053743 0.04335251 0.04335251
#>  [2,] 0.30936530 0.21693850 0.45357116
#>  [3,] 0.08556057 1.03197476 0.66212966
#>  [4,] 0.50799873 0.50799873 0.50799873
#>  [5,] 0.08912087 0.12239261 0.12239261
#>  [6,] 0.02364749 0.02630318 0.02364749
#>  [7,] 0.08192851 0.08192851 0.08192851
#>  [8,] 0.07244109 0.46278155 0.46278155
#>  [9,] 0.05523627 0.00156648 0.18377168
#> [10,] 0.37145728 0.37145728 0.37145728
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>              [,1]        [,2]       [,3]
#>  [1,] 0.265913226 0.105504519 0.10550452
#>  [2,] 0.003557627 0.980617510 0.21738158
#>  [3,] 0.811579335 0.051840086 0.05184009
#>  [4,] 0.526302754 0.596156579 0.42748142
#>  [5,] 0.031713953 0.007917178 0.03171395
#>  [6,] 0.335631138 0.009240500 0.51089721
#>  [7,] 0.144409680 0.683102914 0.15742446
#>  [8,] 0.127688455 0.120986881 0.12446005
#>  [9,] 0.640037414 0.211868210 0.64003741
#> [10,] 0.699165236 0.699165236 0.69916524
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.211028352 0.63783520 0.87148060
#>  [2,] 0.002614876 0.10520937 0.34466938
#>  [3,] 0.152373110 0.18728279 0.07970240
#>  [4,] 0.717620766 0.08130478 0.72633324
#>  [5,] 0.898557210 0.15463847 0.07441545
#>  [6,] 0.456452808 0.14374561 0.45645281
#>  [7,] 0.250649107 0.07942169 0.06300444
#>  [8,] 0.693073393 0.13840745 0.72575452
#>  [9,] 0.162296522 0.16229652 0.30626044
#> [10,] 0.031707364 0.03170736 0.03170736
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.56011028 0.24562914 0.560110280
#>  [2,] 0.07418064 0.45623419 0.084669552
#>  [3,] 0.45771100 0.48722752 0.487227523
#>  [4,] 0.13476453 0.13476453 0.134764532
#>  [5,] 0.43432755 0.43432755 0.008688559
#>  [6,] 0.43169564 1.44674126 0.347380882
#>  [7,] 0.55582177 0.62662704 0.626627037
#>  [8,] 0.16292420 0.19154143 0.402903460
#>  [9,] 0.55742669 0.55742669 0.013467806
#> [10,] 0.10907234 0.04205263 0.079277671

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.06981721 0.04690336 0.34717661
#>  [2,] 0.03953072 0.07158739 0.75811730
#>  [3,] 0.23999359 0.38878997 0.38878997
#>  [4,] 0.09306334 0.09306334 0.09306334
#>  [5,] 0.09450180 0.05008886 0.05008886
#>  [6,] 1.19211085 0.50120950 0.50120950
#>  [7,] 0.48170291 0.62419692 0.70323838
#>  [8,] 1.09038414 0.04364079 1.79684315
#>  [9,] 0.10061002 0.02243314 0.10061002
#> [10,] 0.50782933 0.11841935 0.30225531
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46552641 0.08745799 0.01198273
#>  [2,] 0.47450367 0.24483547 0.47450367
#>  [3,] 0.80392095 0.45722188 0.45722188
#>  [4,] 0.07229862 0.15683273 0.15683273
#>  [5,] 0.46631421 0.46631421 0.46631421
#>  [6,] 0.31747733 0.07722752 0.35311985
#>  [7,] 0.78710710 0.78710710 0.54800510
#>  [8,] 0.90711689 0.49661201 0.64122645
#>  [9,] 0.93086909 0.64813628 0.31466441
#> [10,] 0.13026718 0.13026718 0.13026718
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.61757094 0.38226557 0.20231666
#>  [2,] 0.01110104 0.01110104 0.01110104
#>  [3,] 0.11933756 0.11933756 0.11933756
#>  [4,] 1.38957609 0.10289682 0.47209142
#>  [5,] 0.44477187 0.22086512 0.44477187
#>  [6,] 0.04516823 0.89145298 0.72527047
#>  [7,] 0.19461790 0.33668148 0.33668148
#>  [8,] 0.25816622 0.29956909 0.26767808
#>  [9,] 0.70635465 0.08795733 0.70635465
#> [10,] 0.44317489 0.75147423 0.75147423
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 0.137466448 0.36641468 0.008959449
#>  [2,] 0.437599037 0.43759904 0.437599037
#>  [3,] 0.015132610 0.01513261 0.015132610
#>  [4,] 0.003637841 1.86367530 0.604555185
#>  [5,] 0.124618042 1.06869044 0.124618042
#>  [6,] 0.081570804 0.08157080 0.081570804
#>  [7,] 0.691594275 1.20710570 0.519704155
#>  [8,] 0.305470247 0.30547025 0.305470247
#>  [9,] 0.270277076 0.27274965 0.272749654
#> [10,] 0.055398583 0.17520943 0.173154055