Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]      [,3]
#>  [1,] 1.42666592 1.69387591 0.2111702
#>  [2,] 1.71642026 0.95113323 2.2034065
#>  [3,] 0.02119373 0.09457694 0.7704186
#>  [4,] 3.65745067 1.69969899 2.0993186
#>  [5,] 1.59867831 1.63993286 2.2334451
#>  [6,] 1.25389990 3.55537761 3.4959199
#>  [7,] 0.53539141 0.53539141 0.5353914
#>  [8,] 0.70546433 0.12399087 0.7054643
#>  [9,] 0.33450647 2.49771180 0.6611945
#> [10,] 2.45567251 4.23437757 0.6973475
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.45124036 0.45124036 0.45124036
#>  [2,] 2.07657985 2.07657985 2.07657985
#>  [3,] 1.63022394 1.63022394 1.63022394
#>  [4,] 0.13053260 0.13053260 0.13053260
#>  [5,] 1.39343398 1.39343398 1.39343398
#>  [6,] 0.05983337 0.05983337 0.05983337
#>  [7,] 0.96987002 0.96987002 0.96987002
#>  [8,] 1.70087754 1.70087754 1.70087754
#>  [9,] 0.86017382 0.86017382 0.86017382
#> [10,] 1.19269343 1.19269343 1.19269343
rpextmo(10, 3, b = 1)      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 2.9120792 0.76507913 1.2667107
#>  [2,] 0.6580911 1.11615961 0.6367748
#>  [3,] 2.3519810 0.41192163 0.1492359
#>  [4,] 1.2061122 0.51573313 6.4673595
#>  [5,] 4.4639375 0.29864500 0.6806635
#>  [6,] 0.3738534 0.12574049 1.5272862
#>  [7,] 0.7576504 0.76125216 0.3295232
#>  [8,] 3.6840950 1.46205850 0.8585358
#>  [9,] 0.1574294 0.05049637 1.3524003
#> [10,] 0.5066047 1.09802155 2.0656945

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.08809236 1.60071703 1.6007170
#>  [2,] 0.82090543 2.42425409 2.4242541
#>  [3,] 0.12367590 0.38570708 1.8062914
#>  [4,] 0.49882372 3.21651716 1.1100310
#>  [5,] 0.01481079 2.81660506 0.8959594
#>  [6,] 3.14875105 2.35799542 0.2140089
#>  [7,] 4.34927436 1.48262396 3.1942549
#>  [8,] 1.47516981 0.79849486 1.4751698
#>  [9,] 2.16848649 6.48947865 0.5517596
#> [10,] 4.39141580 0.07302406 0.2701603
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9041178 0.9041178 0.9041178
#>  [2,] 0.3964017 0.3964017 0.3964017
#>  [3,] 0.6199178 0.6199178 0.6199178
#>  [4,] 1.0638878 1.0638878 1.0638878
#>  [5,] 0.5186131 0.5186131 0.5186131
#>  [6,] 0.7302551 0.7302551 0.7302551
#>  [7,] 0.4534904 0.4534904 0.4534904
#>  [8,] 0.4333688 0.4333688 0.4333688
#>  [9,] 0.8740019 0.8740019 0.8740019
#> [10,] 1.4294940 1.4294940 1.4294940
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>               [,1]       [,2]      [,3]
#>  [1,] 0.9835592208 0.63917560 0.8157128
#>  [2,] 0.9114026541 2.11828221 0.6546779
#>  [3,] 0.9451612094 0.58467777 2.1111534
#>  [4,] 0.9152996707 0.16580577 0.7489200
#>  [5,] 0.2888563522 0.15173672 2.7529570
#>  [6,] 1.8302589952 1.89785460 1.2260960
#>  [7,] 0.0006711809 0.34548928 1.2829483
#>  [8,] 2.9172981245 0.01098956 0.6066441
#>  [9,] 0.3893287578 0.87779926 0.1398607
#> [10,] 0.2008315571 1.02595697 0.2504022

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.41927785 3.59099029 0.77612120
#>  [2,] 0.97789190 2.72365041 5.92677319
#>  [3,] 1.22894009 2.03449239 2.03449239
#>  [4,] 0.01026361 0.01026361 0.01026361
#>  [5,] 0.96921232 0.79432368 0.86904323
#>  [6,] 0.24407935 0.43089470 0.87961955
#>  [7,] 1.48349926 3.01053829 1.70101385
#>  [8,] 0.04717872 0.04717872 0.04717872
#>  [9,] 0.03102284 2.54875483 3.07723745
#> [10,] 0.91892296 0.30728557 0.91892296
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5388884 0.5388884 0.5388884
#>  [2,] 0.4905722 0.4905722 0.4905722
#>  [3,] 0.7499633 0.7499633 0.7499633
#>  [4,] 1.5923379 1.5923379 1.5923379
#>  [5,] 0.3797967 0.3797967 0.3797967
#>  [6,] 2.3375714 2.3375714 2.3375714
#>  [7,] 0.1706799 0.1706799 0.1706799
#>  [8,] 0.4252098 0.4252098 0.4252098
#>  [9,] 0.2349854 0.2349854 0.2349854
#> [10,] 0.2754158 0.2754158 0.2754158
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 1.61158169 0.8208228 0.4591234
#>  [2,] 0.01883207 0.2386810 0.5156710
#>  [3,] 1.54970821 0.7151593 1.6780126
#>  [4,] 1.36714044 0.8240567 0.5076225
#>  [5,] 0.48755555 0.8418421 1.6288378
#>  [6,] 0.58038789 1.3317805 0.3316679
#>  [7,] 0.06023969 0.4459011 3.2729710
#>  [8,] 0.36442446 0.7953253 1.5767699
#>  [9,] 0.31091956 0.3767957 2.1802700
#> [10,] 2.73462358 3.7374414 2.5373240

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.84706144 1.8470614 1.8470614
#>  [2,] 0.71050571 2.1611860 2.8558807
#>  [3,] 0.78520062 3.6035558 0.9154244
#>  [4,] 0.12576266 0.9464333 2.2501661
#>  [5,] 1.25844856 3.1652065 1.3947293
#>  [6,] 0.81020073 1.5011429 1.7175944
#>  [7,] 0.09501548 2.2017971 4.0111318
#>  [8,] 2.10082927 0.3596523 1.5237316
#>  [9,] 5.03532862 0.7489042 1.2195247
#> [10,] 1.18312742 2.6001314 1.1571233
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.59198334 1.59198334 1.59198334
#>  [2,] 0.86227871 0.86227871 0.86227871
#>  [3,] 3.32851936 3.32851936 3.32851936
#>  [4,] 1.12774458 1.12774458 1.12774458
#>  [5,] 2.02819845 2.02819845 2.02819845
#>  [6,] 0.10826597 0.10826597 0.10826597
#>  [7,] 2.37843327 2.37843327 2.37843327
#>  [8,] 0.10167801 0.10167801 0.10167801
#>  [9,] 0.02622407 0.02622407 0.02622407
#> [10,] 1.37078284 1.37078284 1.37078284
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.07804459 1.95755213 0.73109532
#>  [2,] 1.13329106 1.20016958 0.41522049
#>  [3,] 1.09073796 0.11725802 0.35958546
#>  [4,] 0.04968660 0.26751706 0.01639175
#>  [5,] 0.33966526 0.60631804 1.08620740
#>  [6,] 1.80115232 0.02779842 0.05322158
#>  [7,] 0.66746444 2.91643691 1.19933000
#>  [8,] 0.05452473 0.43125632 0.12180271
#>  [9,] 0.45178429 0.27976244 0.29926087
#> [10,] 1.81609907 0.03524821 1.50087917

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 3.39197754 4.7425978 1.62095787
#>  [2,] 0.97815278 0.8347603 0.97815278
#>  [3,] 0.06550138 2.8135034 2.36290079
#>  [4,] 1.20654502 1.8142152 1.35850933
#>  [5,] 1.20876553 1.8711785 1.87117852
#>  [6,] 1.12152176 0.3943729 0.97151721
#>  [7,] 0.56408799 1.2402141 1.62314329
#>  [8,] 1.09706022 1.7533641 2.69273935
#>  [9,] 0.89321283 1.0536866 0.03453871
#> [10,] 0.64690963 1.9413125 2.33203989
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.54088040 1.54088040 1.54088040
#>  [2,] 0.04421512 0.04421512 0.04421512
#>  [3,] 1.97375092 1.97375092 1.97375092
#>  [4,] 1.48237665 1.48237665 1.48237665
#>  [5,] 0.26500272 0.26500272 0.26500272
#>  [6,] 0.42678343 0.42678343 0.42678343
#>  [7,] 0.94463284 0.94463284 0.94463284
#>  [8,] 0.15927168 0.15927168 0.15927168
#>  [9,] 0.95549767 0.95549767 0.95549767
#> [10,] 1.17393459 1.17393459 1.17393459
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1005358 0.28000707 0.1095745
#>  [2,] 3.9975620 0.03575073 2.8476887
#>  [3,] 0.5146758 0.92009899 0.3415012
#>  [4,] 6.6526827 0.74512951 1.7531057
#>  [5,] 1.0730904 3.00722408 4.7445868
#>  [6,] 0.1826896 2.68134557 2.9838095
#>  [7,] 0.2104933 1.48595749 0.1209643
#>  [8,] 0.0798463 0.14545125 2.2917765
#>  [9,] 0.3435119 0.56564405 0.9281283
#> [10,] 1.4868834 0.52993033 1.1145964

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0500911 2.5152396 8.4378269
#>  [2,] 3.1649122 0.3654962 1.3665883
#>  [3,] 2.6364037 1.2841456 1.2323096
#>  [4,] 0.4289222 0.4289222 0.4289222
#>  [5,] 0.6079995 1.0214341 1.6368200
#>  [6,] 0.4290825 0.4290825 0.4290825
#>  [7,] 1.1073209 1.1073209 0.2158355
#>  [8,] 0.2710213 0.2710213 0.1164720
#>  [9,] 0.5951561 0.8503411 0.6628041
#> [10,] 1.3003358 2.7918302 2.7918302
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.24829665 1.24829665 1.24829665
#>  [2,] 1.50730430 1.50730430 1.50730430
#>  [3,] 0.03786881 0.03786881 0.03786881
#>  [4,] 0.65880704 0.65880704 0.65880704
#>  [5,] 0.02569442 0.02569442 0.02569442
#>  [6,] 1.06327189 1.06327189 1.06327189
#>  [7,] 2.98046543 2.98046543 2.98046543
#>  [8,] 0.13323368 0.13323368 0.13323368
#>  [9,] 3.18875398 3.18875398 3.18875398
#> [10,] 2.61066004 2.61066004 2.61066004
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]      [,2]        [,3]
#>  [1,] 0.51267250 1.4245818 0.216553354
#>  [2,] 1.51647737 1.1806012 0.773484171
#>  [3,] 0.63960147 2.6772188 0.642818041
#>  [4,] 0.07730006 1.8256703 0.034310856
#>  [5,] 0.30987309 0.1002036 0.192712178
#>  [6,] 1.75044054 0.4433935 0.001427258
#>  [7,] 1.51869630 1.3392892 4.184296109
#>  [8,] 0.47685146 0.8914786 0.139889642
#>  [9,] 1.96040413 2.4693465 0.025200169
#> [10,] 1.76219571 1.1248436 1.321360092

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 3.2202760 3.58826507 2.2883649
#>  [2,] 2.4171441 0.29204066 2.9140886
#>  [3,] 1.3121163 1.59385638 0.2787535
#>  [4,] 1.7643011 1.72547461 3.1595896
#>  [5,] 0.3279002 0.01508704 1.9323538
#>  [6,] 0.4314495 2.18819866 2.1881987
#>  [7,] 2.9911103 2.99111026 0.2324414
#>  [8,] 0.7527786 1.88924374 0.9427397
#>  [9,] 0.1858001 1.00588973 4.4341688
#> [10,] 0.5077754 1.01818735 2.0910346
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.65227122 0.65227122 0.65227122
#>  [2,] 2.77102611 2.77102611 2.77102611
#>  [3,] 0.34850579 0.34850579 0.34850579
#>  [4,] 1.23885351 1.23885351 1.23885351
#>  [5,] 5.33997820 5.33997820 5.33997820
#>  [6,] 0.13920987 0.13920987 0.13920987
#>  [7,] 0.09465253 0.09465253 0.09465253
#>  [8,] 0.08835444 0.08835444 0.08835444
#>  [9,] 1.67295361 1.67295361 1.67295361
#> [10,] 0.11304463 0.11304463 0.11304463
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>              [,1]      [,2]       [,3]
#>  [1,] 0.014026547 0.8118093 0.47512888
#>  [2,] 0.327665406 0.2166204 0.08665240
#>  [3,] 0.002042104 1.4893429 0.64705620
#>  [4,] 0.877436442 0.5435978 0.05867465
#>  [5,] 1.107904323 1.3485888 1.23488037
#>  [6,] 0.398669192 0.8780435 0.33283910
#>  [7,] 1.006487003 2.5393587 0.17422977
#>  [8,] 1.303702868 1.3027522 0.17300677
#>  [9,] 0.610246495 0.6036978 0.18590483
#> [10,] 1.552210225 0.1884679 0.52388251

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 4.0818730 0.67833725 2.8844423
#>  [2,] 1.7077112 7.41445745 1.7597912
#>  [3,] 3.9970712 0.86630952 1.3844680
#>  [4,] 1.1391872 1.59137739 1.7265697
#>  [5,] 0.5795136 1.68539937 1.6002103
#>  [6,] 0.2209794 0.22097938 0.2209794
#>  [7,] 5.7965599 1.76664309 4.0383993
#>  [8,] 3.9331558 0.63973709 1.1329369
#>  [9,] 5.5262271 1.82526186 2.9862106
#> [10,] 3.3462672 0.07744276 1.1529540
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.00975335 0.00975335 0.00975335
#>  [2,] 1.24128314 1.24128314 1.24128314
#>  [3,] 0.23761218 0.23761218 0.23761218
#>  [4,] 0.65372969 0.65372969 0.65372969
#>  [5,] 1.84128343 1.84128343 1.84128343
#>  [6,] 1.59285421 1.59285421 1.59285421
#>  [7,] 0.47695928 0.47695928 0.47695928
#>  [8,] 0.84310829 0.84310829 0.84310829
#>  [9,] 2.59116147 2.59116147 2.59116147
#> [10,] 1.90351632 1.90351632 1.90351632
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.75068453 0.6258392 0.35756085
#>  [2,] 0.18000504 0.4630032 1.29949022
#>  [3,] 2.26896107 3.1980062 0.32262329
#>  [4,] 2.41815226 0.3214125 1.34613263
#>  [5,] 0.03599616 0.6847183 0.53348494
#>  [6,] 1.16348600 1.7432797 1.04063294
#>  [7,] 0.11586653 0.2256526 2.03280039
#>  [8,] 2.47780100 2.3753213 2.00692658
#>  [9,] 1.22044637 1.8414684 0.01441375
#> [10,] 0.19301133 0.6927889 3.48646217

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.383341868 1.3833419 0.9473919
#>  [2,] 4.311027744 1.0104957 1.3213510
#>  [3,] 0.030060242 1.7695455 0.9822331
#>  [4,] 0.881300489 0.4803960 0.9853202
#>  [5,] 0.125127372 0.2052482 0.4021105
#>  [6,] 0.004274612 2.1899087 3.5306259
#>  [7,] 1.290233415 2.6256061 0.2247106
#>  [8,] 2.185515432 3.8003181 0.7803671
#>  [9,] 1.621228434 0.2857216 1.3372064
#> [10,] 1.565106022 0.5534807 1.3776861
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27685395 0.27685395 0.27685395
#>  [2,] 2.21071263 2.21071263 2.21071263
#>  [3,] 0.09451319 0.09451319 0.09451319
#>  [4,] 1.97693654 1.97693654 1.97693654
#>  [5,] 1.35497746 1.35497746 1.35497746
#>  [6,] 0.11895451 0.11895451 0.11895451
#>  [7,] 0.96746841 0.96746841 0.96746841
#>  [8,] 2.33663310 2.33663310 2.33663310
#>  [9,] 0.07779045 0.07779045 0.07779045
#> [10,] 4.38215186 4.38215186 4.38215186
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]        [,2]      [,3]
#>  [1,] 0.7560274 1.672375605 0.2075825
#>  [2,] 0.4552999 1.420924921 0.4716327
#>  [3,] 0.2554645 0.539370263 0.8360228
#>  [4,] 0.3469144 3.465768859 1.0151006
#>  [5,] 2.5065117 0.206762921 0.1071605
#>  [6,] 0.2853615 2.950911656 0.5963122
#>  [7,] 0.1168303 0.579310782 2.3432821
#>  [8,] 0.1580758 0.795331632 0.1641378
#>  [9,] 2.0839495 1.577067989 0.6373628
#> [10,] 0.8177065 0.004589515 0.1161995

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2073005 0.02081079 0.8923877
#>  [2,] 1.6638281 0.72101657 1.4463101
#>  [3,] 1.8352254 0.52807469 0.4603776
#>  [4,] 2.3972548 0.46595783 2.3972548
#>  [5,] 1.5478676 0.95442924 1.4715809
#>  [6,] 0.1733224 0.17332236 0.1733224
#>  [7,] 0.2475099 0.24750987 0.2475099
#>  [8,] 0.9710508 4.88885673 1.2846216
#>  [9,] 0.0635894 0.06358940 0.0635894
#> [10,] 1.5278367 4.18001941 0.4147445
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23909235 0.23909235 0.23909235
#>  [2,] 0.79837096 0.79837096 0.79837096
#>  [3,] 1.88760180 1.88760180 1.88760180
#>  [4,] 1.17673932 1.17673932 1.17673932
#>  [5,] 0.31061620 0.31061620 0.31061620
#>  [6,] 0.09239272 0.09239272 0.09239272
#>  [7,] 1.02608966 1.02608966 1.02608966
#>  [8,] 0.44235516 0.44235516 0.44235516
#>  [9,] 0.61097896 0.61097896 0.61097896
#> [10,] 0.41386336 0.41386336 0.41386336
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26628686 0.18442449 0.06159823
#>  [2,] 0.07406434 0.57531814 0.73536443
#>  [3,] 0.31720830 0.17324606 0.16184813
#>  [4,] 1.78911343 4.18144856 0.04472442
#>  [5,] 1.54710128 1.19311704 0.28018016
#>  [6,] 1.22188190 0.71165490 0.25265872
#>  [7,] 2.30398179 0.09875621 1.39030901
#>  [8,] 0.21985960 0.05779944 0.38777169
#>  [9,] 0.39796797 0.54501141 0.21524262
#> [10,] 0.31546209 1.29579024 1.41748976

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.09773704 1.7589817 2.4475990
#>  [2,] 0.45987243 0.2436599 0.4598724
#>  [3,] 0.35238315 0.3523832 0.4261227
#>  [4,] 0.46915253 0.5047200 0.5047200
#>  [5,] 1.18867361 1.1703444 0.8103320
#>  [6,] 0.39704228 0.9563784 1.1243951
#>  [7,] 0.06567796 1.3700284 1.5253879
#>  [8,] 2.07827047 0.4282618 0.5168097
#>  [9,] 0.56808364 0.5909136 0.6995838
#> [10,] 0.84563731 0.1495629 0.6384243
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36956494 0.36956494 0.17570999
#>  [2,] 2.28345138 0.08528960 1.27070753
#>  [3,] 0.71175313 0.18436604 0.49831315
#>  [4,] 0.42091090 0.49050610 0.49050610
#>  [5,] 0.63339392 0.02833597 0.63339392
#>  [6,] 1.23806196 1.33174002 4.57214652
#>  [7,] 0.94223913 0.01439586 0.04402364
#>  [8,] 0.04390029 0.82431590 0.82431590
#>  [9,] 0.27498963 0.81440890 0.37097620
#> [10,] 0.23688514 0.23688514 0.23688514
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2077198 0.94610960 0.96504189
#>  [2,] 0.3846211 0.38462106 1.36436182
#>  [3,] 0.6120565 0.20524246 1.20741710
#>  [4,] 0.8222680 0.05364487 0.05364487
#>  [5,] 0.1369859 0.36201744 0.13698592
#>  [6,] 0.2530832 0.45830379 0.45830379
#>  [7,] 1.2318060 0.17963785 1.23180595
#>  [8,] 1.2701728 0.29060625 0.14018046
#>  [9,] 0.7953121 0.29054150 1.15703438
#> [10,] 0.2743913 1.06264031 3.25121624
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3294680 0.74047050 0.7404705
#>  [2,] 0.3723185 0.01675061 0.3723185
#>  [3,] 0.1472800 0.85608321 0.8560832
#>  [4,] 1.3835286 0.84886194 0.8488619
#>  [5,] 1.4014235 0.36058151 2.9922139
#>  [6,] 1.9323450 1.93234503 1.1703703
#>  [7,] 0.7738903 1.16345541 1.7897950
#>  [8,] 0.1090674 0.61864256 0.1090674
#>  [9,] 0.3608065 1.27534255 0.1713460
#> [10,] 0.7420269 1.70331608 2.3154942
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03459579 0.03459579 0.03459579
#>  [2,] 0.59422281 0.24033672 0.26301760
#>  [3,] 0.79152703 0.29068795 0.79152703
#>  [4,] 0.62417963 1.20225057 0.41114157
#>  [5,] 1.65543088 1.65975727 1.92304382
#>  [6,] 0.95734496 0.95734496 1.27858946
#>  [7,] 0.63139730 0.17272871 0.12471631
#>  [8,] 1.40913394 0.61519167 2.21635239
#>  [9,] 1.91216300 1.44335042 1.91216300
#> [10,] 0.67353811 0.40821923 0.50700391

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.85472712 1.85472712 1.85472712
#>  [2,] 3.77321889 2.70714324 2.45700435
#>  [3,] 3.30204675 2.19615281 0.04895179
#>  [4,] 1.57347089 1.35395587 0.27631207
#>  [5,] 3.34514070 1.41874352 0.01961153
#>  [6,] 0.08058864 0.05319982 0.08058864
#>  [7,] 0.01431679 0.02410382 0.02410382
#>  [8,] 0.73125973 0.73125973 0.73125973
#>  [9,] 2.32896526 0.04760015 0.22976206
#> [10,] 1.22439856 0.98443710 0.65636895
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 2.3582769 0.45662572 1.0415276
#>  [2,] 0.4806492 2.46306036 5.5315032
#>  [3,] 7.6588816 0.58940874 0.1544606
#>  [4,] 0.3843248 0.09228716 0.2211476
#>  [5,] 0.8734953 4.04090905 0.8734953
#>  [6,] 1.6564407 0.53826347 1.0911770
#>  [7,] 2.0209060 2.94729739 6.7312453
#>  [8,] 1.8001920 2.62559510 0.9244147
#>  [9,] 1.2232761 3.49130434 1.8493151
#> [10,] 0.4532015 0.43101788 0.4532015
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 6.40172337 1.6450661 2.8620734
#>  [2,] 0.11991638 1.3824790 1.3604429
#>  [3,] 0.13574760 0.8692100 0.8692100
#>  [4,] 1.32928238 0.2444959 2.1507282
#>  [5,] 0.01402383 0.2667527 0.2174076
#>  [6,] 0.59100847 0.5910085 0.5910085
#>  [7,] 0.32094966 0.3392251 0.2872698
#>  [8,] 1.35311628 1.0487857 0.4088210
#>  [9,] 0.01412424 3.5851431 1.3708342
#> [10,] 3.75154617 3.7515462 2.1601748
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4250232 1.3931626 1.4250232
#>  [2,] 0.2055151 1.0880852 2.5514665
#>  [3,] 0.2767771 0.6063096 0.2386187
#>  [4,] 0.9383941 2.1208873 1.1398380
#>  [5,] 2.8142687 0.2842725 1.7530814
#>  [6,] 0.6476554 0.1149184 4.0303622
#>  [7,] 1.0955817 0.3030116 0.7679346
#>  [8,] 0.2303246 4.3819995 4.0408947
#>  [9,] 0.1240574 3.1753059 3.1913423
#> [10,] 1.5659098 1.5659098 0.9794389
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 1.283683091 2.23133048 1.35597342
#>  [2,] 0.194864898 1.01585211 2.04802124
#>  [3,] 1.469221685 1.46922169 0.16486438
#>  [4,] 0.806164105 2.66563604 1.40741056
#>  [5,] 0.526808816 0.52680882 0.52680882
#>  [6,] 0.008605986 2.16245637 1.29512595
#>  [7,] 0.480881980 0.48088198 0.48088198
#>  [8,] 0.821559909 0.07163975 0.82155991
#>  [9,] 0.167799392 0.16779939 0.03996773
#> [10,] 0.548404314 0.54840431 0.54840431

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>              [,1]        [,2]        [,3]
#>  [1,] 1.068147738 1.068147738 0.192911559
#>  [2,] 0.148600535 0.148600535 0.462321441
#>  [3,] 0.093635455 0.577596110 0.105180185
#>  [4,] 0.014676108 0.073490835 1.290652428
#>  [5,] 0.611619910 0.020838753 0.611619910
#>  [6,] 0.147442982 0.148836659 0.058197935
#>  [7,] 0.079792774 0.079792774 0.079792774
#>  [8,] 0.003579601 0.003579601 0.003579601
#>  [9,] 0.057331662 0.613823739 1.422286168
#> [10,] 0.051381891 0.038371518 1.086576837
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.66240148 0.639376588 0.462828867
#>  [2,] 1.28873541 0.206870894 0.206870894
#>  [3,] 0.44205440 1.655730966 0.110495962
#>  [4,] 0.09270505 0.058291035 0.058291035
#>  [5,] 0.95326380 0.953263803 0.390007076
#>  [6,] 0.40632006 0.607127858 0.521283871
#>  [7,] 0.24111005 0.241110050 0.241110050
#>  [8,] 1.26839197 0.009483897 0.009483897
#>  [9,] 0.74423073 0.744230730 3.660658175
#> [10,] 0.01640414 0.016404139 0.013803769
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31098303 0.57146053 0.32082134
#>  [2,] 1.09938499 0.34021192 0.34021192
#>  [3,] 0.50598054 0.53141989 0.53965795
#>  [4,] 0.33474779 0.33474779 0.33474779
#>  [5,] 1.22792945 0.37218720 1.22792945
#>  [6,] 0.71785017 0.86278616 0.18487962
#>  [7,] 0.09628722 0.09628722 0.09628722
#>  [8,] 0.71014510 0.71014510 0.67525455
#>  [9,] 0.40647471 0.40647471 0.40647471
#> [10,] 0.22058884 0.22058884 0.22058884
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>              [,1]        [,2]       [,3]
#>  [1,] 1.194454921 0.853261524 1.10264597
#>  [2,] 0.089662769 1.291712606 0.08966277
#>  [3,] 0.383462309 0.383462309 0.38346231
#>  [4,] 0.242155660 0.242155660 0.24215566
#>  [5,] 0.200404241 0.200404241 0.20040424
#>  [6,] 0.442148600 0.048310223 1.70235756
#>  [7,] 0.169688777 0.169688777 0.31691625
#>  [8,] 0.006302003 0.006302003 0.61020574
#>  [9,] 0.312432063 1.870270762 1.10454154
#> [10,] 0.265382112 0.398755070 0.16377293
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27291387 0.57475128 0.27291387
#>  [2,] 0.09240135 0.09240135 0.09240135
#>  [3,] 0.68645632 0.68645632 0.68645632
#>  [4,] 0.22274117 0.22274117 0.22274117
#>  [5,] 0.03605341 0.03605341 0.03292755
#>  [6,] 0.07268407 0.07268407 0.07268407
#>  [7,] 0.62662104 0.05700528 0.05700528
#>  [8,] 0.06837276 0.06837276 0.18974144
#>  [9,] 0.54257137 0.79688391 0.62310300
#> [10,] 0.22052282 0.57626556 0.57626556

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01609328 0.74278331 0.53528992
#>  [2,] 0.06266414 0.06266414 0.06266414
#>  [3,] 0.36718497 0.06874612 0.36718497
#>  [4,] 0.08047204 0.08047204 0.03225827
#>  [5,] 0.11601325 0.12339225 0.12339225
#>  [6,] 0.08625089 0.08625089 0.08625089
#>  [7,] 0.01067345 0.01067345 0.01067345
#>  [8,] 0.32010697 0.16524672 0.98920968
#>  [9,] 0.01261806 0.01261806 0.01261806
#> [10,] 0.71185040 0.16115921 0.59034026
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41677829 0.37295469 0.07167583
#>  [2,] 0.12054133 0.12054133 0.12054133
#>  [3,] 0.02819174 0.10167794 0.19638498
#>  [4,] 0.48939380 0.05757969 0.48317094
#>  [5,] 0.31475101 0.31475101 0.31475101
#>  [6,] 0.22633450 0.35031358 0.35031358
#>  [7,] 0.34842466 1.15730367 0.86574396
#>  [8,] 0.15206705 0.15206705 0.15206705
#>  [9,] 1.15353862 0.15404834 0.74271228
#> [10,] 0.13167163 0.59455632 0.86547760
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02474720 0.02474720 0.02474720
#>  [2,] 0.07235975 0.07235975 0.07235975
#>  [3,] 0.28820724 0.40663015 0.46958445
#>  [4,] 0.28155884 0.28155884 0.28155884
#>  [5,] 1.07930466 0.78223765 0.22315251
#>  [6,] 0.04610897 0.04610897 0.04610897
#>  [7,] 0.58392071 0.28541847 0.16756313
#>  [8,] 0.17128924 0.29030720 0.15375496
#>  [9,] 0.06181610 0.76588345 0.06181610
#> [10,] 0.17460896 0.17460896 0.05431618
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]      [,2]        [,3]
#>  [1,] 0.52484142 0.1055178 0.028842635
#>  [2,] 0.08119196 0.4984369 0.081191961
#>  [3,] 1.15481028 1.1146213 0.002506901
#>  [4,] 1.18159937 0.6204948 0.620494780
#>  [5,] 0.94956088 0.4378773 1.607340416
#>  [6,] 0.12354903 1.3244905 0.200591436
#>  [7,] 0.22682448 0.2268245 0.226824477
#>  [8,] 0.22103015 0.2210302 0.221030154
#>  [9,] 0.51497155 0.5149716 0.346869334
#> [10,] 0.10486339 0.0153361 0.104863390

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04098149 0.98278063 0.98278063
#>  [2,] 0.20702248 0.20702248 0.20702248
#>  [3,] 0.08060936 0.08060936 0.08060936
#>  [4,] 0.37018631 0.39035856 0.21881059
#>  [5,] 0.88284839 0.23479322 0.23479322
#>  [6,] 0.26074996 0.61553835 0.85251060
#>  [7,] 0.24579775 0.24579775 0.24579775
#>  [8,] 0.12218270 0.07047559 0.05603208
#>  [9,] 0.10405045 0.10405045 0.10405045
#> [10,] 0.32248948 0.32248948 0.32248948
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.463855982 0.059692010 0.15071135
#>  [2,] 0.006612399 0.006612399 0.76192535
#>  [3,] 0.553176488 0.080071174 0.17781677
#>  [4,] 0.107507129 0.414757018 0.98181574
#>  [5,] 0.225681326 0.254568839 0.41567445
#>  [6,] 0.547782945 0.640731857 1.60082649
#>  [7,] 0.971336655 0.509926266 0.33828168
#>  [8,] 0.326892280 0.326892280 0.32689228
#>  [9,] 0.545462932 1.252248209 0.09285155
#> [10,] 0.165275082 0.673285405 0.64664326
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11566010 0.02788286 0.02788286
#>  [2,] 1.66432610 0.38646396 1.85182739
#>  [3,] 0.98786461 0.51189649 0.03210430
#>  [4,] 1.06560002 0.50175567 1.06560002
#>  [5,] 0.47296465 0.47296465 0.34802423
#>  [6,] 0.07700058 0.69177890 0.62943755
#>  [7,] 0.17750924 1.19951660 0.46469049
#>  [8,] 0.15650294 0.34296884 0.34296884
#>  [9,] 0.12992225 0.12992225 0.12992225
#> [10,] 0.16419647 0.12184512 0.16419647
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.024963763 0.06539243 0.06539243
#>  [2,] 0.083014309 0.67106900 0.13306711
#>  [3,] 0.190500620 0.38417495 0.19050062
#>  [4,] 0.949501474 0.15147671 0.56142775
#>  [5,] 0.009840427 0.09373587 0.02615764
#>  [6,] 0.220719912 0.26489002 0.13490505
#>  [7,] 0.371485876 0.78484582 0.05783125
#>  [8,] 0.077537594 0.39998497 0.39998497
#>  [9,] 0.577775649 0.57777565 0.57777565
#> [10,] 0.076716489 0.20425068 0.20425068

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]         [,2]       [,3]
#>  [1,] 1.32618397 0.0016936654 0.29369815
#>  [2,] 0.04624680 0.0462467975 0.04624680
#>  [3,] 0.01876168 0.0008878178 0.01876168
#>  [4,] 0.78154673 0.3228880238 1.43344265
#>  [5,] 0.58828009 0.5882800859 1.39927600
#>  [6,] 0.59589631 0.3813669033 0.26426231
#>  [7,] 0.24620449 0.2462044866 0.20604099
#>  [8,] 0.10942899 1.2086162380 0.75753082
#>  [9,] 0.12379561 0.1237956123 0.12379561
#> [10,] 0.21038441 0.2103844131 0.11740694
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0129390 0.57648804 0.34798625
#>  [2,] 0.5872871 0.58728709 0.58728709
#>  [3,] 0.2088852 0.20888519 0.04035020
#>  [4,] 0.2045470 0.60342157 0.19550221
#>  [5,] 0.4568203 0.08264576 0.08264576
#>  [6,] 0.9211732 0.82969470 0.67693733
#>  [7,] 0.2132653 0.21326532 0.21326532
#>  [8,] 0.3953439 0.43158318 0.06595426
#>  [9,] 0.4057872 0.40578723 0.45257832
#> [10,] 0.2876576 0.45620890 1.45744378
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02155392 0.07064774 0.07064774
#>  [2,] 0.29139557 0.47672369 0.47672369
#>  [3,] 0.28375136 0.28375136 0.28375136
#>  [4,] 0.40927539 0.21678931 0.80902401
#>  [5,] 0.03364707 0.04566657 0.03364707
#>  [6,] 0.30228876 0.26860585 0.30975658
#>  [7,] 0.03240974 0.34417090 0.34417090
#>  [8,] 0.32889888 0.38379420 0.38379420
#>  [9,] 0.26092725 0.26092725 0.53179810
#> [10,] 0.15750821 0.11265464 0.11265464
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01378454 0.03900567 0.01378454
#>  [2,] 0.24069701 0.12388050 0.06486274
#>  [3,] 0.11058045 0.58448180 0.15763295
#>  [4,] 0.95396749 0.30380052 0.30380052
#>  [5,] 0.14340152 0.14340152 0.14340152
#>  [6,] 0.29592263 0.15418216 0.33583345
#>  [7,] 0.29108937 0.27005666 0.29108937
#>  [8,] 0.43137330 0.03525938 0.03525938
#>  [9,] 0.12912051 0.05319724 0.05319724
#> [10,] 0.12336606 0.12336606 0.13857707