Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0916828 1.1163334 2.0916828
#>  [2,] 1.6738987 0.1397851 0.2270680
#>  [3,] 0.4294707 0.9550772 1.6237034
#>  [4,] 1.9589590 0.5689106 1.0437129
#>  [5,] 5.9033690 3.2447433 0.7490945
#>  [6,] 1.1281268 3.2085917 1.0082156
#>  [7,] 0.4027071 2.6838920 0.5609438
#>  [8,] 0.7424086 0.5067686 0.7424086
#>  [9,] 1.0590542 3.5493856 1.7791041
#> [10,] 1.9082645 0.5723544 2.4833748
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2421530 1.2421530 1.2421530
#>  [2,] 0.3694418 0.3694418 0.3694418
#>  [3,] 0.9247548 0.9247548 0.9247548
#>  [4,] 0.5964056 0.5964056 0.5964056
#>  [5,] 0.8713085 0.8713085 0.8713085
#>  [6,] 1.6932172 1.6932172 1.6932172
#>  [7,] 1.5015110 1.5015110 1.5015110
#>  [8,] 2.6016673 2.6016673 2.6016673
#>  [9,] 0.5744470 0.5744470 0.5744470
#> [10,] 0.4455906 0.4455906 0.4455906
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17785167 0.33062497 1.23237858
#>  [2,] 0.50936753 0.33350668 0.90596114
#>  [3,] 0.25989078 2.25025943 0.02913345
#>  [4,] 1.09961296 0.93674015 2.55633141
#>  [5,] 0.86927636 1.71482324 0.89120756
#>  [6,] 0.32890346 0.82867021 3.65125809
#>  [7,] 1.45358660 0.19448630 0.99215683
#>  [8,] 0.06807478 0.08719275 1.41073723
#>  [9,] 2.57701793 0.21807325 0.42533575
#> [10,] 0.18820786 2.74326444 0.35854001

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.7197582 0.55558781 1.15241510
#>  [2,] 0.9069848 1.19117278 2.37888626
#>  [3,] 0.5556683 0.05771274 2.61216682
#>  [4,] 2.4705762 0.04072265 2.47057619
#>  [5,] 1.8065645 0.06396983 1.78696055
#>  [6,] 1.6547583 1.65475830 0.23787011
#>  [7,] 0.5380232 1.25086198 1.41504848
#>  [8,] 0.2891585 1.20929874 1.20929874
#>  [9,] 0.3392880 0.18276088 0.06520079
#> [10,] 1.1589486 1.15894860 0.45401212
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04078559 0.04078559 0.04078559
#>  [2,] 0.11958032 0.11958032 0.11958032
#>  [3,] 0.74920983 0.74920983 0.74920983
#>  [4,] 0.47991918 0.47991918 0.47991918
#>  [5,] 1.85414182 1.85414182 1.85414182
#>  [6,] 0.05945864 0.05945864 0.05945864
#>  [7,] 1.02889663 1.02889663 1.02889663
#>  [8,] 0.18036215 0.18036215 0.18036215
#>  [9,] 1.21064727 1.21064727 1.21064727
#> [10,] 3.32661542 3.32661542 3.32661542
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3827407 1.6620084 2.93112691
#>  [2,] 2.4908300 0.1133089 1.18367902
#>  [3,] 0.1792549 3.4814994 1.41338813
#>  [4,] 0.5271342 0.1414734 0.17952283
#>  [5,] 0.4567373 2.3503152 0.45387939
#>  [6,] 1.8273519 1.8797972 0.89030306
#>  [7,] 1.4620240 0.1244491 3.90185887
#>  [8,] 2.3074624 2.3063470 0.07841753
#>  [9,] 2.1190940 0.5018193 0.05201169
#> [10,] 2.8681846 2.2202364 3.05811320

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01884692 0.01884692 0.01884692
#>  [2,] 0.55491976 3.57172702 2.19862079
#>  [3,] 0.83231463 0.16861119 0.68889445
#>  [4,] 0.34604626 0.77186710 0.55876469
#>  [5,] 1.33943059 1.33943059 1.33943059
#>  [6,] 0.67453824 7.24885311 4.09249284
#>  [7,] 1.12317935 1.55842779 5.47875098
#>  [8,] 0.28699566 0.28699566 0.28699566
#>  [9,] 0.16977115 0.38748781 0.38748781
#> [10,] 0.70990445 0.70990445 0.59818597
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9300436 1.9300436 1.9300436
#>  [2,] 2.1692669 2.1692669 2.1692669
#>  [3,] 0.3044626 0.3044626 0.3044626
#>  [4,] 0.7502310 0.7502310 0.7502310
#>  [5,] 0.1327146 0.1327146 0.1327146
#>  [6,] 0.4569014 0.4569014 0.4569014
#>  [7,] 0.2925763 0.2925763 0.2925763
#>  [8,] 0.5314412 0.5314412 0.5314412
#>  [9,] 0.8972513 0.8972513 0.8972513
#> [10,] 1.1255563 1.1255563 1.1255563
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.37999835 1.4496340 0.73217336
#>  [2,] 0.60447312 1.5359031 1.82290856
#>  [3,] 0.07689363 0.2139502 1.52853220
#>  [4,] 0.92921348 1.5020608 0.14046681
#>  [5,] 0.76577874 0.4681298 0.09787641
#>  [6,] 0.72791369 0.8686383 1.53097151
#>  [7,] 0.59486528 1.1206587 1.56370621
#>  [8,] 2.05995088 1.1317538 0.51518123
#>  [9,] 0.99220690 1.2523661 0.60697327
#> [10,] 0.11341347 1.1270111 0.61208082

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.77020549 2.00021797 5.5960187
#>  [2,] 0.20736388 0.20736388 0.2073639
#>  [3,] 1.18624079 0.66251931 0.8798897
#>  [4,] 1.01604455 1.01604455 0.3762010
#>  [5,] 0.38138180 0.38138180 0.3813818
#>  [6,] 0.01030850 5.66079812 0.6258238
#>  [7,] 0.99876136 0.05264681 1.0562104
#>  [8,] 0.06114929 2.23265219 0.9040641
#>  [9,] 0.79783522 0.46061209 1.5579393
#> [10,] 1.14603330 0.86242339 1.4697551
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.88305106 1.88305106 1.88305106
#>  [2,] 1.01829635 1.01829635 1.01829635
#>  [3,] 0.09393902 0.09393902 0.09393902
#>  [4,] 0.08153939 0.08153939 0.08153939
#>  [5,] 1.46181695 1.46181695 1.46181695
#>  [6,] 0.61868328 0.61868328 0.61868328
#>  [7,] 0.71022356 0.71022356 0.71022356
#>  [8,] 0.85459220 0.85459220 0.85459220
#>  [9,] 0.63532879 0.63532879 0.63532879
#> [10,] 0.13117482 0.13117482 0.13117482
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]        [,3]
#>  [1,] 1.57894275 2.13336904 0.641729997
#>  [2,] 0.07345329 0.04990326 1.744412424
#>  [3,] 0.17135416 0.40020893 1.383519390
#>  [4,] 0.78877450 0.06920686 4.126830085
#>  [5,] 0.16123244 0.53811454 0.403798178
#>  [6,] 0.14543445 3.46336982 2.797651414
#>  [7,] 1.57110075 0.87947850 0.874523186
#>  [8,] 0.27341789 0.53015529 0.926817583
#>  [9,] 0.33273732 0.19149781 0.077069984
#> [10,] 2.85303551 0.09585916 0.002535823

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.09895120 3.91932033 3.2165555
#>  [2,] 3.03949435 0.39730878 1.3408666
#>  [3,] 2.76093183 5.52998803 3.0734181
#>  [4,] 0.02097318 0.04211926 0.4503812
#>  [5,] 4.56290139 1.49241886 2.6557316
#>  [6,] 0.60930288 1.51862016 1.5186202
#>  [7,] 0.58422760 0.02629872 2.6799132
#>  [8,] 0.96346871 0.91983876 0.9634687
#>  [9,] 0.18105470 0.23407973 0.4830839
#> [10,] 2.16975155 3.75533903 3.9201864
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8161206 0.8161206 0.8161206
#>  [2,] 2.2998797 2.2998797 2.2998797
#>  [3,] 1.0958560 1.0958560 1.0958560
#>  [4,] 0.5458453 0.5458453 0.5458453
#>  [5,] 0.3440183 0.3440183 0.3440183
#>  [6,] 0.2228657 0.2228657 0.2228657
#>  [7,] 2.6628379 2.6628379 2.6628379
#>  [8,] 3.3065704 3.3065704 3.3065704
#>  [9,] 0.1489756 0.1489756 0.1489756
#> [10,] 2.0804185 2.0804185 2.0804185
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.60655296 0.5260551 0.9587512
#>  [2,] 0.08039680 0.1604679 1.0744829
#>  [3,] 0.88968953 0.1147380 1.3593549
#>  [4,] 0.80302923 1.1609431 0.3205474
#>  [5,] 0.33194334 0.2786664 0.1905650
#>  [6,] 0.48808942 0.1354884 0.1624842
#>  [7,] 0.07031483 1.0458914 1.3835107
#>  [8,] 0.73655307 1.0632100 0.8400155
#>  [9,] 0.31483433 1.0384710 0.9082307
#> [10,] 0.28261916 2.9639079 0.6514703

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.92798816 1.4683519 0.3415539
#>  [2,] 2.72019091 1.5706107 1.6431632
#>  [3,] 0.05152810 0.0515281 0.0515281
#>  [4,] 1.70818382 0.4710577 2.1221617
#>  [5,] 2.86064980 0.2766167 1.7839788
#>  [6,] 0.09833943 2.8096569 5.0973207
#>  [7,] 4.11738858 3.0167666 3.7868111
#>  [8,] 0.36466129 0.6863361 0.3625807
#>  [9,] 0.63492938 0.9676195 4.4086048
#> [10,] 0.04884385 2.7043587 1.0374982
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.19948665 2.19948665 2.19948665
#>  [2,] 0.15798243 0.15798243 0.15798243
#>  [3,] 0.66063441 0.66063441 0.66063441
#>  [4,] 2.73319174 2.73319174 2.73319174
#>  [5,] 0.69177963 0.69177963 0.69177963
#>  [6,] 0.06351744 0.06351744 0.06351744
#>  [7,] 0.64290221 0.64290221 0.64290221
#>  [8,] 0.90456945 0.90456945 0.90456945
#>  [9,] 0.09250343 0.09250343 0.09250343
#> [10,] 0.67976358 0.67976358 0.67976358
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.34179698 0.4464332 0.02608651
#>  [2,] 0.45537637 1.2101298 0.10762035
#>  [3,] 1.35954673 2.3526774 0.23226747
#>  [4,] 0.14739265 1.0175640 0.27330256
#>  [5,] 0.86125173 2.7961527 3.20758197
#>  [6,] 0.89572647 0.2138989 0.55762174
#>  [7,] 0.75813438 1.6445576 1.10765780
#>  [8,] 1.37402441 0.9468827 1.05667222
#>  [9,] 0.01800726 2.4833103 0.58087314
#> [10,] 0.34907303 0.2773104 1.33233240

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>               [,1]       [,2]      [,3]
#>  [1,]  0.168584703 2.05911863 2.0591186
#>  [2,]  0.136294063 0.33065291 0.3306529
#>  [3,]  4.691781696 2.77201338 0.3406816
#>  [4,]  0.537264082 0.53726408 0.5372641
#>  [5,]  0.494099341 1.33077666 2.8563345
#>  [6,]  0.101935298 2.46128462 3.8878793
#>  [7,]  0.006342754 0.04371499 0.5646028
#>  [8,]  1.839495650 0.04750605 1.7157748
#>  [9,] 10.186724398 2.86120669 1.6370734
#> [10,]  1.489622606 0.52256831 1.0105783
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.51212832 0.51212832 0.51212832
#>  [2,] 3.63563822 3.63563822 3.63563822
#>  [3,] 0.46392745 0.46392745 0.46392745
#>  [4,] 0.35915648 0.35915648 0.35915648
#>  [5,] 0.04926916 0.04926916 0.04926916
#>  [6,] 0.08918802 0.08918802 0.08918802
#>  [7,] 0.09852734 0.09852734 0.09852734
#>  [8,] 0.53896765 0.53896765 0.53896765
#>  [9,] 2.80472607 2.80472607 2.80472607
#> [10,] 0.28568025 0.28568025 0.28568025
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.39014873 1.0489924 0.24496497
#>  [2,] 1.03132176 0.5528163 0.10304038
#>  [3,] 0.06328227 1.1208385 0.04460773
#>  [4,] 0.44036455 0.2899308 0.90945886
#>  [5,] 2.25916625 0.2716642 1.14569295
#>  [6,] 1.83669755 1.1763615 2.83985257
#>  [7,] 2.35314514 0.1241905 0.86147782
#>  [8,] 0.95099694 0.3697580 0.20240152
#>  [9,] 0.41521428 2.0164603 0.08387350
#> [10,] 1.34814454 0.5985532 0.89016441

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 2.9363868 1.49708171 0.5768485
#>  [2,] 0.9213736 2.50126358 3.2638298
#>  [3,] 0.2619978 0.16978202 0.1627679
#>  [4,] 1.0154531 0.39163483 0.7888504
#>  [5,] 1.0140812 0.27376523 1.0140812
#>  [6,] 1.1583730 0.01422874 0.3448769
#>  [7,] 3.1043242 0.05826769 1.2640834
#>  [8,] 1.5331594 1.49594704 0.4651366
#>  [9,] 2.3792736 0.63862675 2.7724572
#> [10,] 2.4163749 0.01441665 0.5991223
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.19714699 2.19714699 2.19714699
#>  [2,] 2.14173994 2.14173994 2.14173994
#>  [3,] 0.44022073 0.44022073 0.44022073
#>  [4,] 0.60720883 0.60720883 0.60720883
#>  [5,] 0.08534547 0.08534547 0.08534547
#>  [6,] 0.03293652 0.03293652 0.03293652
#>  [7,] 1.39732053 1.39732053 1.39732053
#>  [8,] 2.05879727 2.05879727 2.05879727
#>  [9,] 1.61825033 1.61825033 1.61825033
#> [10,] 0.38540009 0.38540009 0.38540009
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3924662 1.57738601 0.2716444
#>  [2,] 1.5773426 0.05620161 0.1859730
#>  [3,] 0.1827284 0.85816612 0.7645601
#>  [4,] 0.1274404 0.48707103 0.1534382
#>  [5,] 1.2652775 0.40491912 1.2612120
#>  [6,] 0.4436614 1.43275101 0.3183597
#>  [7,] 0.8812001 1.82088822 0.3197596
#>  [8,] 1.7861358 0.05512328 1.9833256
#>  [9,] 6.4214586 2.24122857 0.7870011
#> [10,] 2.3853651 0.90403799 2.4147930

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.04159622 4.09590882 1.059161046
#>  [2,] 1.00753092 0.71536432 1.158875746
#>  [3,] 0.35865657 0.49498287 0.003004534
#>  [4,] 0.12681990 1.85123260 8.678053812
#>  [5,] 2.07143762 0.96110068 0.180027345
#>  [6,] 1.66502610 0.09652262 0.929258095
#>  [7,] 2.12702496 0.33830299 1.042134668
#>  [8,] 1.10221765 0.06279077 0.093590757
#>  [9,] 3.12198483 0.20194214 1.808701881
#> [10,] 1.68550742 3.40346412 2.857334776
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34062531 0.34062531 0.34062531
#>  [2,] 0.32105600 0.32105600 0.32105600
#>  [3,] 2.73650988 2.73650988 2.73650988
#>  [4,] 5.68059896 5.68059896 5.68059896
#>  [5,] 0.42419532 0.42419532 0.42419532
#>  [6,] 0.60589693 0.60589693 0.60589693
#>  [7,] 1.83558866 1.83558866 1.83558866
#>  [8,] 0.71788853 0.71788853 0.71788853
#>  [9,] 0.90137670 0.90137670 0.90137670
#> [10,] 0.07544114 0.07544114 0.07544114
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1929833 1.68463198 0.17153803
#>  [2,] 0.5853547 0.73008268 1.62495574
#>  [3,] 1.6959321 0.52222524 0.02267144
#>  [4,] 0.1208235 0.09190404 0.69560191
#>  [5,] 1.1726179 0.19947219 1.40976944
#>  [6,] 2.3576830 0.14205019 1.15783457
#>  [7,] 0.3754325 0.35482708 0.95104763
#>  [8,] 0.6981699 0.52862756 1.06489275
#>  [9,] 0.4178060 1.30517702 0.13060447
#> [10,] 0.9046151 0.41942799 0.71990054

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0735487 2.8821067 1.3282866
#>  [2,] 0.1006143 0.1006143 0.1006143
#>  [3,] 4.1928165 4.6154541 2.8754164
#>  [4,] 0.5330680 1.3779458 0.5496765
#>  [5,] 0.9096907 0.9096907 0.9096907
#>  [6,] 4.4138489 2.0742799 0.2406193
#>  [7,] 0.2851728 1.9447409 0.1373552
#>  [8,] 0.5276523 0.5276523 0.5276523
#>  [9,] 0.5225747 0.4190630 0.9719846
#> [10,] 0.2961279 3.1950277 1.8750822
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.69531170 3.69531170 3.69531170
#>  [2,] 0.02516323 0.02516323 0.02516323
#>  [3,] 0.54271378 0.54271378 0.54271378
#>  [4,] 0.18104307 0.18104307 0.18104307
#>  [5,] 0.26631709 0.26631709 0.26631709
#>  [6,] 0.55380335 0.55380335 0.55380335
#>  [7,] 0.22770946 0.22770946 0.22770946
#>  [8,] 0.06138540 0.06138540 0.06138540
#>  [9,] 0.63563258 0.63563258 0.63563258
#> [10,] 0.96404962 0.96404962 0.96404962
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8661730 1.1472060 0.53255524
#>  [2,] 0.6119559 1.3859472 0.57248101
#>  [3,] 0.1855906 0.9547554 1.66930674
#>  [4,] 1.0621454 2.6907088 1.69334331
#>  [5,] 1.6944177 2.8018803 0.14177110
#>  [6,] 0.1342963 0.1828630 1.12300151
#>  [7,] 0.3599739 0.1100040 1.08677532
#>  [8,] 0.1027085 0.6038492 0.08124178
#>  [9,] 0.1541394 0.1937703 1.14777963
#> [10,] 0.9101546 1.1382982 2.46067990

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 1.11466788 1.06520404 1.932437719
#>  [2,] 0.81976034 0.57486765 1.723177205
#>  [3,] 0.74666151 0.74666151 0.746661513
#>  [4,] 0.06514332 0.03127207 0.008208173
#>  [5,] 1.69937022 2.35969992 1.324032454
#>  [6,] 2.07291127 0.46216318 0.462163178
#>  [7,] 0.66702747 0.33046987 1.071166630
#>  [8,] 4.30642230 0.52814588 0.273396671
#>  [9,] 0.99407943 0.80450940 0.016859380
#> [10,] 0.18613269 1.24526213 0.278234373
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2259231 0.68074191 0.68074191
#>  [2,] 0.5282519 0.48855077 0.61461181
#>  [3,] 0.1659096 0.16590955 0.04809506
#>  [4,] 0.6886910 0.52379843 0.29825219
#>  [5,] 0.2594423 0.57541225 0.25944233
#>  [6,] 3.2471506 0.03104733 0.07347702
#>  [7,] 0.1750110 0.17501102 0.17501102
#>  [8,] 2.2802399 0.08749259 0.51894758
#>  [9,] 2.2209656 0.09333390 0.09445048
#> [10,] 0.8341556 1.94343706 0.83415563
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>               [,1]      [,2]        [,3]
#>  [1,] 0.2253574911 0.3776026 0.225357491
#>  [2,] 1.0770184239 0.2068477 1.077018424
#>  [3,] 0.6931587188 1.6199713 0.099353541
#>  [4,] 0.3779857562 0.5441087 1.197917790
#>  [5,] 0.1222504126 3.0059683 0.003505084
#>  [6,] 0.3348941896 0.2895852 1.024675979
#>  [7,] 0.2825227210 0.2825227 0.681738767
#>  [8,] 1.0635846128 0.1472262 4.012704052
#>  [9,] 0.0008584617 1.0086184 0.620966779
#> [10,] 0.2003299196 0.2003299 1.807961038
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.53909462 1.04765934 0.337402930
#>  [2,] 0.33534225 0.52232206 0.151552659
#>  [3,] 1.18345863 1.17457267 1.537547464
#>  [4,] 0.02630314 0.27498950 0.274989499
#>  [5,] 0.93665008 0.34024891 1.685387634
#>  [6,] 0.30966284 0.66865275 0.008192238
#>  [7,] 0.07017792 0.07017792 0.572382620
#>  [8,] 0.68640570 0.46871877 1.629699041
#>  [9,] 0.47167056 0.65378690 0.471670556
#> [10,] 0.16080151 0.16080151 0.015963836
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19180128 1.06154234 0.81063952
#>  [2,] 0.07104630 0.07104630 0.07104630
#>  [3,] 1.70211152 0.61045691 1.10068302
#>  [4,] 0.30773851 0.30773851 0.32939381
#>  [5,] 0.34641737 0.01109962 0.01109962
#>  [6,] 1.37730761 2.04529569 2.17971983
#>  [7,] 0.55174364 0.18631821 0.34956226
#>  [8,] 1.91924724 0.48918080 0.20585542
#>  [9,] 0.09934363 0.93500626 0.34616968
#> [10,] 0.69223184 0.69284974 0.23037181

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 2.91446150 1.1114978 2.6059169
#>  [2,] 1.79335599 0.3807395 0.6952869
#>  [3,] 0.57816206 2.9535347 1.7586939
#>  [4,] 1.82626810 4.9086382 0.2274354
#>  [5,] 2.77759619 0.4135634 0.2098609
#>  [6,] 0.34912645 0.2859991 0.3491265
#>  [7,] 1.57109472 0.1448290 2.6326519
#>  [8,] 0.42946921 0.4294692 0.4294692
#>  [9,] 0.81090979 2.5164097 2.2754955
#> [10,] 0.04041116 2.2544284 0.4352020
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3514513 0.44807846 0.4480785
#>  [2,] 0.8171513 0.24846694 1.4265929
#>  [3,] 1.1552925 1.03487610 0.7260340
#>  [4,] 6.5849078 0.23678675 2.0456397
#>  [5,] 2.9054469 2.66460212 2.9054469
#>  [6,] 1.6191210 0.36347584 1.6191210
#>  [7,] 1.5104655 0.12622163 0.3485221
#>  [8,] 0.2146127 0.07012482 3.4826400
#>  [9,] 1.4528932 0.87331226 0.1566519
#> [10,] 0.3516953 0.35169531 0.3516953
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.56399667 0.56399667 0.5639967
#>  [2,] 1.24417118 4.22384777 0.1055258
#>  [3,] 1.72737922 1.95130890 0.5405384
#>  [4,] 0.05572371 0.73532352 0.7353235
#>  [5,] 1.32134864 0.73688544 0.7168737
#>  [6,] 0.87216673 1.09694826 0.9258846
#>  [7,] 0.85494106 0.85494106 0.8549411
#>  [8,] 0.36425017 2.48004035 1.2292310
#>  [9,] 0.27039525 0.34375863 0.3437586
#> [10,] 1.52104766 0.07105562 2.4914583
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.2101168 0.69606382 0.8939892
#>  [2,] 1.2792755 0.12982279 0.7022336
#>  [3,] 0.5654523 2.39164376 0.8753928
#>  [4,] 2.4517058 2.45170580 2.4517058
#>  [5,] 1.4609601 1.46096006 1.4609601
#>  [6,] 2.5200888 0.09902662 1.4017588
#>  [7,] 0.9404176 3.20183665 0.4264753
#>  [8,] 2.0149704 0.85416437 1.7460017
#>  [9,] 2.7669451 1.79360250 3.0069555
#> [10,] 8.7789592 0.64258611 1.9457603
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 1.0471561 1.522750690 1.22642044
#>  [2,] 1.2984805 0.658400137 0.01290472
#>  [3,] 5.7168842 5.716884180 1.49005025
#>  [4,] 1.1497274 2.289708310 1.28757712
#>  [5,] 2.4916156 4.491657402 1.68308197
#>  [6,] 0.3199821 0.319982086 0.16696335
#>  [7,] 0.5677156 0.004933662 0.01428549
#>  [8,] 1.2340569 1.828824454 0.88790598
#>  [9,] 1.9881732 0.892267184 1.31352321
#> [10,] 2.0855330 0.425267677 0.99686626

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06673972 0.02985195 0.06673972
#>  [2,] 0.29587173 0.33026178 0.29587173
#>  [3,] 0.96750114 0.96750114 1.84402568
#>  [4,] 0.82343640 0.05899706 0.44621872
#>  [5,] 0.04511882 1.64553770 0.04511882
#>  [6,] 0.37844424 0.70879937 0.70879937
#>  [7,] 0.19165920 0.40593294 0.40593294
#>  [8,] 0.21234376 0.08467276 0.21234376
#>  [9,] 0.45036796 0.45036796 0.45036796
#> [10,] 0.02898600 0.02898600 0.02898600
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.66942038 0.66942038 0.16413856
#>  [2,] 0.27957929 0.27957929 0.30295276
#>  [3,] 0.08395357 0.08395357 0.08395357
#>  [4,] 0.01386961 0.01386961 0.01386961
#>  [5,] 0.19984325 0.19984325 0.19984325
#>  [6,] 0.04221643 0.04221643 0.04221643
#>  [7,] 0.38910843 0.47276742 0.33754958
#>  [8,] 0.51716865 0.21258876 1.05335616
#>  [9,] 0.32354365 0.36453671 0.36453671
#> [10,] 0.01183950 0.01183950 0.01183950
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.13118128 0.50572373 1.13118128
#>  [2,] 0.21744202 1.56821951 0.87282001
#>  [3,] 0.03210285 0.03210285 0.03210285
#>  [4,] 0.07359392 0.07359392 0.07359392
#>  [5,] 0.75064926 0.75064926 0.75064926
#>  [6,] 0.33452105 0.33452105 0.33452105
#>  [7,] 0.14570714 1.84680529 0.14570714
#>  [8,] 0.14981073 0.14981073 0.38899544
#>  [9,] 0.62190542 0.62190542 0.62190542
#> [10,] 2.25355256 1.36760699 0.78156364
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.329403821 0.26662181 0.09412194
#>  [2,] 0.149915355 0.14991536 0.14991536
#>  [3,] 0.006573968 0.74854301 0.74854301
#>  [4,] 0.472532245 0.08627988 0.36029965
#>  [5,] 0.751461524 0.75146152 0.75146152
#>  [6,] 0.460503919 0.46050392 0.36675016
#>  [7,] 0.116712324 0.11671232 0.11671232
#>  [8,] 0.391428982 0.28708377 0.28708377
#>  [9,] 0.126045598 0.12604560 0.12604560
#> [10,] 2.117531392 0.06509011 0.07492611
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>               [,1]       [,2]       [,3]
#>  [1,] 0.7782033429 0.77820334 0.77820334
#>  [2,] 0.2426646212 0.24266462 0.24266462
#>  [3,] 0.0407970134 0.13076523 0.03334184
#>  [4,] 0.2823531200 0.42302841 0.35726707
#>  [5,] 0.2447961649 0.24479616 0.24479616
#>  [6,] 0.5289894314 0.52898943 0.52898943
#>  [7,] 0.1235417908 0.12354179 0.12354179
#>  [8,] 0.0004893294 0.56880857 0.56880857
#>  [9,] 0.3073302721 0.91007760 0.91007760
#> [10,] 0.0960291180 0.09602912 0.07863771

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.61058221 0.61058221 0.49963030
#>  [2,] 0.11371503 0.18624583 0.11371503
#>  [3,] 0.22568501 0.22568501 0.22568501
#>  [4,] 0.17062304 0.05741945 0.17062304
#>  [5,] 0.75889681 0.01143980 0.42464398
#>  [6,] 0.12703325 0.12703325 0.12703325
#>  [7,] 0.03532858 0.03532858 0.03532858
#>  [8,] 0.39471614 1.40048683 1.38422218
#>  [9,] 0.17120035 0.17120035 0.17120035
#> [10,] 0.32934527 0.34297310 0.09238417
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04035624 0.04035624 0.04035624
#>  [2,] 0.16108655 0.09915519 0.69070029
#>  [3,] 0.28478641 0.57450298 0.18128625
#>  [4,] 0.28981641 0.16105185 0.15698660
#>  [5,] 0.66528814 1.27343164 0.66528814
#>  [6,] 0.13540531 0.40318009 0.12978042
#>  [7,] 0.03743746 0.03743746 0.03743746
#>  [8,] 1.18912192 0.46345305 1.03377521
#>  [9,] 0.31998334 0.44485298 0.09504070
#> [10,] 0.25546205 0.24292162 0.24292162
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 2.24463055 2.1641125 2.61922219
#>  [2,] 0.86666112 1.0687647 0.40691283
#>  [3,] 0.12596021 0.1259602 0.22358925
#>  [4,] 0.11188144 0.1118814 0.11188144
#>  [5,] 0.66381954 0.1276182 0.29795502
#>  [6,] 0.41592918 0.3528600 0.75999315
#>  [7,] 0.55567716 0.3399594 0.14982859
#>  [8,] 0.01171006 0.1394955 0.17704334
#>  [9,] 1.03312777 0.1198819 0.08791407
#> [10,] 0.50185137 0.5018514 0.50185137
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.52427099 0.10934593 0.52427099
#>  [2,] 0.16081231 0.16081231 0.16081231
#>  [3,] 0.39908286 0.39908286 0.39908286
#>  [4,] 0.61041781 0.52490314 0.09278742
#>  [5,] 0.05584053 0.53327056 0.53327056
#>  [6,] 0.91649529 0.06996213 0.18508934
#>  [7,] 0.01890655 1.34034188 0.35936937
#>  [8,] 0.46359045 0.31890086 0.80823117
#>  [9,] 0.61839394 0.44460058 0.58280605
#> [10,] 0.79651330 0.46036694 0.79651330

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.19713690 0.59310391 0.1864986
#>  [2,] 0.06415294 0.99669583 0.7637318
#>  [3,] 0.19177357 0.19177357 0.1917736
#>  [4,] 0.46550270 0.21761508 0.1451197
#>  [5,] 0.37018335 0.87261584 0.2551320
#>  [6,] 0.94838510 0.25648399 0.9617462
#>  [7,] 0.32731933 0.04827975 0.3273193
#>  [8,] 0.54630181 0.29150074 1.9267518
#>  [9,] 0.50919013 0.50919013 0.9344414
#> [10,] 0.20792722 0.83498150 1.1211004
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]         [,2]       [,3]
#>  [1,] 0.76823192 0.9154291694 0.89663100
#>  [2,] 0.24820442 0.3484386513 0.39883766
#>  [3,] 0.28209120 0.2820912010 0.28209120
#>  [4,] 0.36852371 0.3685237059 0.16383913
#>  [5,] 0.03074396 0.8679661276 0.07169855
#>  [6,] 0.52046652 0.2765538721 0.86639973
#>  [7,] 0.12340886 0.0005485343 0.80254647
#>  [8,] 0.80419902 0.1427606967 1.03473013
#>  [9,] 0.57302662 0.3240485448 0.57302662
#> [10,] 0.26623270 0.5350907836 0.08587111
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.16364962 1.74095976 1.16655740
#>  [2,] 0.07692390 0.30495982 0.20542953
#>  [3,] 0.10354004 0.07372583 0.04059079
#>  [4,] 0.24342601 0.32292222 0.24342601
#>  [5,] 0.25667698 0.25667698 0.58311704
#>  [6,] 0.14362635 0.55077730 0.57262430
#>  [7,] 0.39463767 0.33716629 1.61633167
#>  [8,] 0.04897219 0.38028268 0.24809956
#>  [9,] 0.04384605 0.50566451 1.16207096
#> [10,] 0.50144282 1.19882835 0.61296931
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.40534182 0.06530427 0.2357648
#>  [2,] 0.31494752 0.17028627 0.6285607
#>  [3,] 0.29617942 0.33173334 0.4695039
#>  [4,] 1.26842458 0.10587658 0.1058766
#>  [5,] 0.18964033 0.60299698 0.6029970
#>  [6,] 0.37786220 0.46959583 0.3778622
#>  [7,] 0.08075612 1.65972947 0.2832961
#>  [8,] 0.99376695 0.29951137 0.1590689
#>  [9,] 0.13549174 0.13549174 0.4871051
#> [10,] 0.32739535 0.32739535 0.5995146

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05198844 0.05716716 0.08951714
#>  [2,] 1.83385021 0.17895439 0.50955209
#>  [3,] 0.58826434 0.58826434 0.06930939
#>  [4,] 0.86076589 0.26795331 0.86076589
#>  [5,] 0.16185526 0.16185526 0.16185526
#>  [6,] 0.23633336 0.08442777 0.23633336
#>  [7,] 0.01462187 0.23513782 0.44563319
#>  [8,] 0.54015166 0.11964122 0.54015166
#>  [9,] 0.13037940 0.17273205 0.17273205
#> [10,] 0.33767073 0.49694042 0.29347893
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>               [,1]       [,2]       [,3]
#>  [1,] 0.2884881951 0.28848820 0.28848820
#>  [2,] 0.4886638936 0.11801443 1.26950716
#>  [3,] 1.0691672766 0.71400175 0.08708851
#>  [4,] 0.2905706967 0.18178696 0.29057070
#>  [5,] 0.3811153049 0.38111530 0.38111530
#>  [6,] 0.8441225574 0.15851673 0.15851673
#>  [7,] 0.1048931242 0.46076471 0.39771831
#>  [8,] 0.7202999369 0.08648321 0.26041912
#>  [9,] 0.3518999635 0.54324446 0.53266690
#> [10,] 0.0008628318 0.10982477 0.10982477
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]        [,2]        [,3]
#>  [1,] 0.56845250 0.278399600 0.278399600
#>  [2,] 0.10667912 0.106679120 0.106679120
#>  [3,] 0.07785595 0.072498356 0.121166981
#>  [4,] 0.82230034 0.032968116 1.068941373
#>  [5,] 0.04445795 0.005006652 0.477565166
#>  [6,] 0.24557335 0.164818745 0.459584234
#>  [7,] 0.53361006 0.420451368 0.536921128
#>  [8,] 0.59123015 0.007525867 0.007525867
#>  [9,] 0.20667417 0.530789245 0.206674174
#> [10,] 0.09198321 0.689080624 0.084126300
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23587836 0.04994063 0.23587836
#>  [2,] 0.35217011 0.35217011 0.35217011
#>  [3,] 0.08639997 0.08639997 0.04939398
#>  [4,] 0.22902958 0.21494220 0.21494220
#>  [5,] 0.74055624 0.34059383 0.17616960
#>  [6,] 0.41755219 1.14166353 1.14166353
#>  [7,] 1.08703733 0.29372791 0.79677656
#>  [8,] 0.80991981 0.31991896 0.06982926
#>  [9,] 0.85532602 0.51287743 0.85532602
#> [10,] 0.19370759 0.11753591 0.19370759