Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01866099 0.08478868 0.34505462
#>  [2,] 1.61186020 0.88806733 3.37801468
#>  [3,] 2.81358939 0.10840299 1.35098872
#>  [4,] 1.96726926 1.96726926 1.12595941
#>  [5,] 0.01395206 0.01395206 0.01395206
#>  [6,] 0.73571497 0.43852538 0.94766076
#>  [7,] 1.76370897 1.33973927 0.03962021
#>  [8,] 0.08483517 0.08483517 0.08483517
#>  [9,] 0.08121376 0.87922570 0.87922570
#> [10,] 4.50166265 3.44072715 1.82569715
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.39158717 1.39158717 1.39158717
#>  [2,] 0.92678622 0.92678622 0.92678622
#>  [3,] 1.70094825 1.70094825 1.70094825
#>  [4,] 0.03890648 0.03890648 0.03890648
#>  [5,] 0.75880693 0.75880693 0.75880693
#>  [6,] 1.80079711 1.80079711 1.80079711
#>  [7,] 1.42359630 1.42359630 1.42359630
#>  [8,] 0.15008629 0.15008629 0.15008629
#>  [9,] 0.29919340 0.29919340 0.29919340
#> [10,] 3.53039875 3.53039875 3.53039875
## independence
rpextmo(10, 3, b = 1)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.95224351 2.8295053 1.0356981
#>  [2,] 2.85225894 1.1268573 0.5837284
#>  [3,] 0.18333445 0.1122367 0.1794377
#>  [4,] 1.04110780 0.8519810 1.1957578
#>  [5,] 0.83927309 0.4200341 0.1974925
#>  [6,] 1.50393316 1.1838515 0.3056272
#>  [7,] 0.08094599 0.4940007 1.7313263
#>  [8,] 0.62851970 1.5228925 0.3770616
#>  [9,] 0.23711422 0.3942152 0.7833212
#> [10,] 1.01722982 0.1703046 0.2963610

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.05693545 1.0576268 1.0576268
#>  [2,] 1.57949147 1.5794915 1.5794915
#>  [3,] 2.05218004 2.8025352 1.5967363
#>  [4,] 2.69931570 1.2254488 0.2556626
#>  [5,] 2.43776032 4.2083030 2.0570257
#>  [6,] 0.68642051 2.5277651 3.0982834
#>  [7,] 1.04808638 0.2654900 1.0480864
#>  [8,] 0.52040421 0.5099089 3.8626555
#>  [9,] 0.01001797 2.0350258 0.1494021
#> [10,] 0.58587376 1.1521253 0.1355240
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28697187 0.28697187 0.28697187
#>  [2,] 4.72419170 4.72419170 4.72419170
#>  [3,] 0.91647743 0.91647743 0.91647743
#>  [4,] 0.14824431 0.14824431 0.14824431
#>  [5,] 0.06310064 0.06310064 0.06310064
#>  [6,] 1.43940680 1.43940680 1.43940680
#>  [7,] 0.05235014 0.05235014 0.05235014
#>  [8,] 0.91763496 0.91763496 0.91763496
#>  [9,] 0.58123669 0.58123669 0.58123669
#> [10,] 0.22728921 0.22728921 0.22728921
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 2.22672522 1.4569209 1.24662960
#>  [2,] 2.30904311 1.7160727 0.10710161
#>  [3,] 1.44679900 1.8996976 0.22110251
#>  [4,] 0.09288283 2.8134513 2.28681532
#>  [5,] 0.59363086 0.7910714 0.42670048
#>  [6,] 2.38381246 3.1735527 0.37328251
#>  [7,] 1.43523184 0.6000176 1.00085592
#>  [8,] 0.61929106 1.1025049 0.44672152
#>  [9,] 1.35230383 1.3039275 0.06349159
#> [10,] 0.67075187 0.3497718 0.05236714

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.54866330 2.8638879 1.0976932
#>  [2,] 3.12960709 0.8658253 0.4537780
#>  [3,] 0.09699357 1.2503389 1.9498519
#>  [4,] 0.06583782 1.4932460 2.1361299
#>  [5,] 0.15294340 3.1592542 1.3032429
#>  [6,] 0.06458751 6.5453245 0.4845501
#>  [7,] 0.15320347 0.3897055 0.5095786
#>  [8,] 0.06178320 1.0973936 2.4471616
#>  [9,] 1.01039786 2.1095059 2.1095059
#> [10,] 0.54607576 2.5796991 0.8585280
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29058375 0.29058375 0.29058375
#>  [2,] 0.04300285 0.04300285 0.04300285
#>  [3,] 0.10254529 0.10254529 0.10254529
#>  [4,] 3.22402242 3.22402242 3.22402242
#>  [5,] 0.10018944 0.10018944 0.10018944
#>  [6,] 0.42695333 0.42695333 0.42695333
#>  [7,] 2.60336204 2.60336204 2.60336204
#>  [8,] 3.61582457 3.61582457 3.61582457
#>  [9,] 0.09853321 0.09853321 0.09853321
#> [10,] 0.23912588 0.23912588 0.23912588
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.4587094 2.3095700 0.41206294
#>  [2,] 1.0791256 0.2000956 1.15176229
#>  [3,] 0.3376648 1.0056648 1.09151351
#>  [4,] 1.8173200 0.7675359 0.08803752
#>  [5,] 0.3638919 0.7096876 0.91160743
#>  [6,] 0.1288619 0.8671994 2.25124276
#>  [7,] 1.0342836 0.0412185 0.70975016
#>  [8,] 0.2616270 1.6205345 2.98077623
#>  [9,] 1.2512128 0.1620208 2.21118571
#> [10,] 0.1180182 0.1179381 0.78330044

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 2.41883904 1.4232652 0.3702613
#>  [2,] 0.98921064 0.9892106 0.1817613
#>  [3,] 0.07048086 0.7472906 0.2572940
#>  [4,] 1.61632868 1.6163287 0.5071925
#>  [5,] 0.75891169 0.7589117 0.7589117
#>  [6,] 0.33859381 0.6059750 0.0138113
#>  [7,] 2.62920225 1.3789405 0.7181889
#>  [8,] 4.57413654 0.1956564 2.5092878
#>  [9,] 0.05348257 0.2303249 0.2303249
#> [10,] 0.46246028 1.2130863 5.0142131
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2417537 0.2417537 0.2417537
#>  [2,] 0.4741595 0.4741595 0.4741595
#>  [3,] 2.5046904 2.5046904 2.5046904
#>  [4,] 0.1377471 0.1377471 0.1377471
#>  [5,] 0.1547744 0.1547744 0.1547744
#>  [6,] 0.5417775 0.5417775 0.5417775
#>  [7,] 1.2898959 1.2898959 1.2898959
#>  [8,] 1.4143850 1.4143850 1.4143850
#>  [9,] 2.8202682 2.8202682 2.8202682
#> [10,] 1.1633982 1.1633982 1.1633982
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2052643 0.59022771 0.6857535
#>  [2,] 0.6691379 0.80508799 0.6949107
#>  [3,] 3.3505076 1.17869251 0.2940379
#>  [4,] 0.5875417 1.87300622 0.6081790
#>  [5,] 4.9481925 0.50606098 0.9291516
#>  [6,] 1.5293096 0.02573068 0.9439439
#>  [7,] 1.1363342 0.06533791 0.9935377
#>  [8,] 1.6805931 0.21602641 0.6248575
#>  [9,] 0.3207993 0.51269938 0.8181797
#> [10,] 0.4424754 0.72776463 1.4317600

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7375209 0.7102808 1.3796462
#>  [2,] 1.8492925 1.3776431 1.6194308
#>  [3,] 2.4689849 2.4689849 1.0972625
#>  [4,] 0.4183385 0.4443444 0.4443444
#>  [5,] 3.8143080 0.9456117 1.3807480
#>  [6,] 3.3008480 1.9516993 0.5933764
#>  [7,] 2.0825471 2.2698361 0.3464191
#>  [8,] 0.5692153 0.2587152 0.5692153
#>  [9,] 1.6770885 0.4986522 0.5117712
#> [10,] 2.2718182 1.4775867 0.2304022
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08696942 0.08696942 0.08696942
#>  [2,] 1.60035903 1.60035903 1.60035903
#>  [3,] 0.25045600 0.25045600 0.25045600
#>  [4,] 0.50002416 0.50002416 0.50002416
#>  [5,] 0.70955422 0.70955422 0.70955422
#>  [6,] 3.01481810 3.01481810 3.01481810
#>  [7,] 1.40994049 1.40994049 1.40994049
#>  [8,] 3.53459911 3.53459911 3.53459911
#>  [9,] 1.21488690 1.21488690 1.21488690
#> [10,] 0.63344589 0.63344589 0.63344589
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1270747 0.03679806 0.51483055
#>  [2,] 0.3482853 2.21647413 0.04629703
#>  [3,] 1.2626412 0.01018199 1.58929358
#>  [4,] 0.1450210 3.30138783 7.92094555
#>  [5,] 0.8410477 1.25246031 3.58885050
#>  [6,] 0.5547510 0.08981342 1.72385555
#>  [7,] 0.9905810 0.34919045 1.09897649
#>  [8,] 0.9766904 0.79754265 0.51066900
#>  [9,] 0.4031559 1.78991722 1.03455487
#> [10,] 0.2466400 0.53096256 0.69163542

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1577008 0.7536813 0.7536813
#>  [2,] 1.1446787 1.1213732 1.1306646
#>  [3,] 2.6500472 1.0813083 1.5399735
#>  [4,] 1.6070937 0.2587031 0.7984633
#>  [5,] 0.9656308 1.0571686 1.0571686
#>  [6,] 3.2194148 1.1809042 0.4940547
#>  [7,] 0.1102865 0.5971713 0.1923165
#>  [8,] 0.5475875 1.1280574 1.1280574
#>  [9,] 1.1206437 1.4461306 0.2150660
#> [10,] 1.4135631 0.2174868 1.9608470
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26027423 0.26027423 0.26027423
#>  [2,] 0.35784394 0.35784394 0.35784394
#>  [3,] 0.14722124 0.14722124 0.14722124
#>  [4,] 0.06321138 0.06321138 0.06321138
#>  [5,] 1.03919144 1.03919144 1.03919144
#>  [6,] 0.35333352 0.35333352 0.35333352
#>  [7,] 0.83027861 0.83027861 0.83027861
#>  [8,] 2.79822023 2.79822023 2.79822023
#>  [9,] 0.88552504 0.88552504 0.88552504
#> [10,] 0.99510218 0.99510218 0.99510218
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.4365331 0.37592656 2.20975649
#>  [2,] 0.3224523 1.05382179 0.54727074
#>  [3,] 0.3981187 0.02825116 0.73166046
#>  [4,] 0.3809327 0.17217913 1.88823107
#>  [5,] 0.2949028 0.72182881 4.45358628
#>  [6,] 0.2704240 3.36552956 0.06945234
#>  [7,] 0.1283106 0.01404160 0.66356251
#>  [8,] 0.5179236 1.40021512 0.18867295
#>  [9,] 2.2037704 0.55773691 1.49687258
#> [10,] 0.9433991 0.94540127 0.41278307

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0785306 1.0802278 0.09007144
#>  [2,] 4.6365005 0.2339384 1.07492882
#>  [3,] 2.5232488 0.2691460 0.61727417
#>  [4,] 3.8385788 0.4081725 1.29105708
#>  [5,] 2.1810587 0.2182089 0.52658041
#>  [6,] 3.7214687 3.2350271 3.72146869
#>  [7,] 6.6297859 0.7980867 0.51641354
#>  [8,] 1.6035703 6.0555582 3.20780823
#>  [9,] 1.1582824 0.4995307 0.61547612
#> [10,] 0.1773864 0.2815015 1.51294261
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.76586073 0.76586073 0.76586073
#>  [2,] 0.98995422 0.98995422 0.98995422
#>  [3,] 0.92936019 0.92936019 0.92936019
#>  [4,] 0.14172703 0.14172703 0.14172703
#>  [5,] 0.15058737 0.15058737 0.15058737
#>  [6,] 1.38343486 1.38343486 1.38343486
#>  [7,] 0.62246368 0.62246368 0.62246368
#>  [8,] 0.30421990 0.30421990 0.30421990
#>  [9,] 0.04536151 0.04536151 0.04536151
#> [10,] 1.25034224 1.25034224 1.25034224
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.57757622 2.4822176 0.15040679
#>  [2,] 1.39006908 0.1076211 0.48030329
#>  [3,] 0.23647768 0.3367305 0.90461169
#>  [4,] 0.71333261 5.0422600 0.99165106
#>  [5,] 0.08721471 0.1140090 1.18026909
#>  [6,] 1.42152816 0.6090064 3.97509070
#>  [7,] 0.78440889 0.7561192 0.49290663
#>  [8,] 0.24693333 3.8355977 0.07779858
#>  [9,] 2.81498345 2.7966723 0.06432344
#> [10,] 1.68399199 1.2455944 0.21769601

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2497395 0.07900555 0.5900570
#>  [2,] 1.8382013 1.83820135 1.0014904
#>  [3,] 2.8910718 0.23856081 2.2547544
#>  [4,] 1.4105950 1.21449166 2.8318239
#>  [5,] 0.0370677 2.12075867 8.0613708
#>  [6,] 0.7994576 2.53971941 4.9829787
#>  [7,] 4.2488280 3.59537927 0.5916014
#>  [8,] 0.5047536 0.37388033 4.8747841
#>  [9,] 5.5933580 0.38213263 0.3502669
#> [10,] 2.3435169 2.34351687 0.4431827
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.49615208 0.49615208 0.49615208
#>  [2,] 2.77213940 2.77213940 2.77213940
#>  [3,] 0.04653134 0.04653134 0.04653134
#>  [4,] 3.21065186 3.21065186 3.21065186
#>  [5,] 2.31955287 2.31955287 2.31955287
#>  [6,] 0.63793751 0.63793751 0.63793751
#>  [7,] 0.11009436 0.11009436 0.11009436
#>  [8,] 0.36946528 0.36946528 0.36946528
#>  [9,] 0.02877847 0.02877847 0.02877847
#> [10,] 0.59526946 0.59526946 0.59526946
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5346178 0.6067912 3.07590064
#>  [2,] 0.1144926 3.3824465 0.15503660
#>  [3,] 1.3388578 3.7214425 0.81436192
#>  [4,] 2.9574439 1.2227238 0.83133831
#>  [5,] 0.5144285 1.0735962 1.09720024
#>  [6,] 0.6002250 0.5275170 2.76346185
#>  [7,] 0.4574835 1.1001802 1.59288290
#>  [8,] 2.0681695 0.5555884 0.43661945
#>  [9,] 0.3935588 0.5075208 0.24196165
#> [10,] 0.1516011 3.8406722 0.02340085

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.37369271 1.6788727 0.3838238
#>  [2,] 3.85950666 3.8595067 0.9097222
#>  [3,] 1.59867871 1.0277390 3.7299326
#>  [4,] 0.11363401 1.5174920 0.2345461
#>  [5,] 0.03087081 0.3587474 2.5238665
#>  [6,] 2.60761360 0.6006125 1.6121269
#>  [7,] 0.03974517 1.0440956 4.8900317
#>  [8,] 5.75228508 0.8694362 1.7711668
#>  [9,] 0.20823473 0.2082347 0.2082347
#> [10,] 1.92999288 0.7383158 1.2293564
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07592387 0.07592387 0.07592387
#>  [2,] 1.48530488 1.48530488 1.48530488
#>  [3,] 0.99385671 0.99385671 0.99385671
#>  [4,] 0.07689907 0.07689907 0.07689907
#>  [5,] 0.03451048 0.03451048 0.03451048
#>  [6,] 0.57287437 0.57287437 0.57287437
#>  [7,] 0.45116121 0.45116121 0.45116121
#>  [8,] 0.05404566 0.05404566 0.05404566
#>  [9,] 0.58568231 0.58568231 0.58568231
#> [10,] 2.54479818 2.54479818 2.54479818
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.17734483 0.6050522 1.5852841
#>  [2,] 0.18687109 0.1096635 0.1422823
#>  [3,] 2.87999653 1.8614107 0.1832798
#>  [4,] 0.18649178 0.3283049 0.5629172
#>  [5,] 2.52954935 0.5397557 0.9862765
#>  [6,] 0.15916663 0.1437100 0.4570709
#>  [7,] 0.45159751 1.7228172 3.2252958
#>  [8,] 0.25097878 0.1498642 0.9102246
#>  [9,] 0.20603374 1.6355472 1.2721597
#> [10,] 0.04824073 0.8461999 0.2734531

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 1.802026323 0.57100292 0.56206116
#>  [2,] 0.537351427 3.67523293 1.18743794
#>  [3,] 3.333415473 0.74080017 2.43323802
#>  [4,] 2.002141027 0.05221737 1.55764510
#>  [5,] 0.234015260 2.90066764 1.00945573
#>  [6,] 1.591825331 2.00560740 2.00560740
#>  [7,] 0.138535463 1.23560048 1.48540834
#>  [8,] 2.794736097 2.73684221 1.43268802
#>  [9,] 2.160952534 5.67329344 0.07869714
#> [10,] 0.008095532 1.37765851 3.65955752
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11114716 0.11114716 0.11114716
#>  [2,] 0.22923699 0.22923699 0.22923699
#>  [3,] 0.06448208 0.06448208 0.06448208
#>  [4,] 0.17626825 0.17626825 0.17626825
#>  [5,] 0.13522672 0.13522672 0.13522672
#>  [6,] 0.56260649 0.56260649 0.56260649
#>  [7,] 1.12501327 1.12501327 1.12501327
#>  [8,] 1.14022111 1.14022111 1.14022111
#>  [9,] 0.55681844 0.55681844 0.55681844
#> [10,] 1.27742562 1.27742562 1.27742562
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.83208416 0.33351987 2.5050979
#>  [2,] 0.07017313 0.67994846 0.7686589
#>  [3,] 0.48358132 1.90267470 2.4955684
#>  [4,] 1.99016345 0.07291105 3.0954853
#>  [5,] 2.87034115 1.44961563 0.2392701
#>  [6,] 0.52199774 0.02522864 1.8904974
#>  [7,] 2.89677655 1.30060159 0.2821898
#>  [8,] 0.96317725 0.16983091 1.3745131
#>  [9,] 3.15553810 0.08848311 0.3288236
#> [10,] 1.99774553 0.38404063 0.2470503

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7642968 0.2043312 1.3275241
#>  [2,] 1.1497783 0.7099280 0.6698433
#>  [3,] 0.1490711 0.7330948 0.1490711
#>  [4,] 2.3327571 0.2335464 0.2335464
#>  [5,] 0.5516042 0.5518001 0.2932410
#>  [6,] 1.8782572 0.6109744 0.6109744
#>  [7,] 0.3178930 0.5923287 0.5630438
#>  [8,] 0.6120559 0.6120559 0.6120559
#>  [9,] 0.1876266 0.2785963 0.1660095
#> [10,] 0.4319697 0.8953299 0.8953299
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.35955408 0.75182674 0.35955408
#>  [2,] 1.08482638 1.01062795 0.19472224
#>  [3,] 0.12533857 0.17782561 0.17782561
#>  [4,] 1.16965165 1.99767318 2.30096549
#>  [5,] 0.28628776 0.28628776 0.27494605
#>  [6,] 0.32520740 0.08946967 0.38848046
#>  [7,] 1.15719148 0.06828646 0.06828646
#>  [8,] 1.91529662 0.62129504 0.58068696
#>  [9,] 0.63764961 0.81323484 0.11441874
#> [10,] 0.08513147 0.08513147 0.66787102
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>              [,1]       [,2]      [,3]
#>  [1,] 1.488220412 1.06040570 0.4746150
#>  [2,] 0.790740825 0.51621515 0.8791285
#>  [3,] 1.022397854 0.04309391 0.2465573
#>  [4,] 0.009132628 0.14794588 0.1478315
#>  [5,] 0.454072655 0.45407266 0.4540727
#>  [6,] 0.366973347 0.36697335 0.3669733
#>  [7,] 0.464402414 0.46440241 0.4644024
#>  [8,] 0.506166882 0.03140135 0.5061669
#>  [9,] 0.903540918 1.08253611 1.0825361
#> [10,] 0.242243508 1.30113266 0.8340281
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.4467471 0.98630716 0.9863072
#>  [2,] 0.5458486 0.54584857 0.5458486
#>  [3,] 0.1954165 0.04649094 1.0930038
#>  [4,] 0.4909998 0.60388757 0.5333058
#>  [5,] 0.5300995 3.04654719 0.2139092
#>  [6,] 0.1063073 0.10630725 0.1063073
#>  [7,] 1.3745279 1.25332772 0.2675832
#>  [8,] 1.4449892 0.49376697 0.8667088
#>  [9,] 0.8765086 0.54132221 0.8765086
#> [10,] 0.4688008 2.74192061 1.2836497
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.21592314 0.21592314 0.2159231
#>  [2,] 0.56154553 0.77761356 0.2601151
#>  [3,] 0.02814336 0.02814336 0.0124498
#>  [4,] 0.49449369 1.26312818 0.4944937
#>  [5,] 0.33905813 1.65854617 2.1416741
#>  [6,] 1.25353666 1.20358719 0.5152518
#>  [7,] 0.40794856 0.40794856 0.7320617
#>  [8,] 0.04637121 0.04637121 1.5951529
#>  [9,] 0.19726530 1.68028385 2.5211370
#> [10,] 0.65918005 0.11476806 1.3065701

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4479192 0.44791923 0.44791923
#>  [2,] 0.7823503 0.71071887 2.24481657
#>  [3,] 0.2935615 1.05209031 0.36294406
#>  [4,] 2.6598966 0.04969001 1.25923748
#>  [5,] 4.3093314 0.52791199 1.57146437
#>  [6,] 1.8242905 0.67826354 0.04519338
#>  [7,] 0.4222395 0.42223954 0.42223954
#>  [8,] 6.1731180 1.83268738 6.76622082
#>  [9,] 0.1247472 0.12474721 0.12474721
#> [10,] 2.5533571 1.78748300 3.41035376
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 2.47338967 3.8637414 2.2544048
#>  [2,] 5.91851455 1.1825116 0.2868595
#>  [3,] 0.70874974 4.1171946 4.1171946
#>  [4,] 0.58330787 1.2605265 1.2605265
#>  [5,] 0.78799238 1.7428580 0.2058708
#>  [6,] 0.15122694 0.2245426 0.2245426
#>  [7,] 0.09343131 1.5469108 0.1513394
#>  [8,] 0.13223213 0.1043824 0.0936859
#>  [9,] 0.35952998 0.7561350 2.2384876
#> [10,] 2.70491177 0.4732213 1.1443128
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.01421198 0.9534520 1.51323715
#>  [2,] 0.81405967 0.1980552 0.27140405
#>  [3,] 0.97782720 1.3718938 0.18210587
#>  [4,] 0.06725918 0.5224277 0.52242769
#>  [5,] 1.70995824 1.7099582 0.04693734
#>  [6,] 5.09746254 0.4966000 0.91027252
#>  [7,] 0.44904487 0.4490449 0.44904487
#>  [8,] 1.38440860 1.3844086 1.38440860
#>  [9,] 0.76288506 0.7705634 0.77056345
#> [10,] 0.47732983 2.8110384 2.81103845
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.39122159 3.07577830 2.1575185
#>  [2,] 0.52901536 4.48259970 0.5328492
#>  [3,] 0.67328766 2.61495156 2.8746616
#>  [4,] 1.49492676 2.85079123 1.4385760
#>  [5,] 2.04530047 0.49457579 0.3236008
#>  [6,] 1.05184667 2.81826119 0.3709435
#>  [7,] 0.37090263 0.05547879 0.3709026
#>  [8,] 1.72028614 0.80485618 1.4140491
#>  [9,] 0.07315986 0.91917267 0.1206429
#> [10,] 0.51480115 1.08890116 0.4325392
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>              [,1]      [,2]       [,3]
#>  [1,] 3.051703287 2.3209304 1.37282787
#>  [2,] 1.185639645 1.1856396 0.99461736
#>  [3,] 0.471260232 0.3520256 0.36982222
#>  [4,] 0.098415383 1.1813146 1.05095525
#>  [5,] 1.192098484 1.9102364 1.24947201
#>  [6,] 3.408156839 1.6122001 1.70066992
#>  [7,] 0.569666845 0.5696668 0.56966684
#>  [8,] 2.110923166 0.4847567 0.24586437
#>  [9,] 0.336273807 0.5309027 2.89036207
#> [10,] 0.001820072 1.6415866 0.09418252

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.44519274 0.44519274 0.44519274
#>  [2,] 0.53011852 0.53011852 0.27308299
#>  [3,] 0.62824681 0.62824681 0.60728232
#>  [4,] 0.75536323 0.75536323 0.66093579
#>  [5,] 0.19507802 0.19507802 0.19507802
#>  [6,] 0.06964464 0.06964464 0.06964464
#>  [7,] 0.68166365 0.68166365 0.68166365
#>  [8,] 0.60864857 0.13345576 0.60864857
#>  [9,] 0.79872331 0.40889597 0.27969691
#> [10,] 0.04066980 0.04066980 0.04066980
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05137511 0.05137511 0.05137511
#>  [2,] 0.48866186 0.48866186 0.48866186
#>  [3,] 0.24070555 0.93907413 0.93907413
#>  [4,] 0.07637066 0.24911664 0.75596144
#>  [5,] 0.03639657 0.03639657 0.03639657
#>  [6,] 0.14567030 0.14567030 0.14567030
#>  [7,] 1.26342282 1.26342282 1.24199296
#>  [8,] 1.38345369 1.72606836 1.81155161
#>  [9,] 0.31270109 0.72079626 1.29869856
#> [10,] 0.71810502 0.57165803 0.71810502
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1310409 0.89999461 1.1450379
#>  [2,] 0.3541694 0.23172624 0.2317262
#>  [3,] 0.5051984 0.32153005 0.3333979
#>  [4,] 1.6796191 1.02993383 1.7513977
#>  [5,] 0.3153576 0.03363303 0.3153576
#>  [6,] 0.6511093 0.19880623 0.6605316
#>  [7,] 0.1931369 0.19313690 0.1931369
#>  [8,] 1.0868418 1.08684176 0.8448659
#>  [9,] 3.6020920 0.51849914 3.4720367
#> [10,] 0.3263196 0.32631957 0.3263196
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.29612908 1.24841071 0.29612908
#>  [2,] 0.01723684 0.01723684 0.01723684
#>  [3,] 0.03779016 0.03779016 0.03779016
#>  [4,] 0.54953207 1.57093164 0.51888589
#>  [5,] 0.36906565 0.43670822 0.43670822
#>  [6,] 0.20801114 0.20801114 0.20801114
#>  [7,] 0.72773317 0.72773317 0.72773317
#>  [8,] 0.28837824 1.05762496 0.68359181
#>  [9,] 0.65515692 1.08297411 1.94605275
#> [10,] 0.89681116 0.37157307 0.89681116
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42028053 0.41131369 0.39407562
#>  [2,] 0.17756058 0.17756058 0.17756058
#>  [3,] 0.04113788 0.04113788 0.24770892
#>  [4,] 0.33897364 0.33897364 0.31149895
#>  [5,] 0.04043911 0.04043911 0.04043911
#>  [6,] 1.49551959 1.04107925 0.41129490
#>  [7,] 1.27836765 1.27836765 1.27836765
#>  [8,] 0.82562771 0.82562771 0.40271847
#>  [9,] 0.34238471 0.34238471 0.34238471
#> [10,] 0.06418026 0.06418026 0.06418026

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.923594802 1.235151959 1.206615575
#>  [2,] 0.332726313 0.332726313 0.533839612
#>  [3,] 0.004151706 0.004151706 0.004151706
#>  [4,] 0.233001404 0.034510468 0.302632077
#>  [5,] 0.042886634 0.042886634 0.042886634
#>  [6,] 0.114530038 0.114530038 0.114530038
#>  [7,] 0.245538740 0.090046089 0.244455648
#>  [8,] 0.136399350 0.136399350 0.313968710
#>  [9,] 0.026333909 0.862174678 0.129419684
#> [10,] 1.116891628 0.066325352 1.116891628
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34099081 0.45174849 0.45174849
#>  [2,] 0.06809761 0.06809761 0.06809761
#>  [3,] 0.44774438 0.31270891 0.44774438
#>  [4,] 0.11083386 0.30649481 0.32448148
#>  [5,] 0.30997032 0.56265428 1.60853287
#>  [6,] 0.22130525 0.18593035 0.22130525
#>  [7,] 0.73223616 0.20317187 0.73223616
#>  [8,] 0.46459514 0.54473685 0.29371680
#>  [9,] 0.23208993 0.03089564 0.03089564
#> [10,] 0.38697887 0.65349885 1.06888279
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.00311837 0.50824267 0.00311837
#>  [2,] 0.01883306 0.34250236 0.34250236
#>  [3,] 0.01540736 0.01540736 0.01540736
#>  [4,] 0.24156460 0.10470126 0.20576709
#>  [5,] 0.02810374 0.02810374 0.02810374
#>  [6,] 0.09761896 0.03319042 0.03319042
#>  [7,] 0.14288557 0.14288557 0.14288557
#>  [8,] 0.57322432 0.57322432 0.57322432
#>  [9,] 0.39872268 0.05551663 0.72869079
#> [10,] 0.07856814 0.07856814 1.37068497
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.72785632 0.06422626 0.10435022
#>  [2,] 0.53453184 0.33170562 0.60211354
#>  [3,] 0.36075805 0.47601703 0.36631708
#>  [4,] 0.21289365 0.63399200 1.29480605
#>  [5,] 0.34133554 0.51139029 1.51484503
#>  [6,] 0.17255903 0.29607933 0.19865977
#>  [7,] 0.18456213 0.47359464 0.47359464
#>  [8,] 0.03969592 0.01049661 0.08056514
#>  [9,] 0.44347850 0.44347850 0.16095149
#> [10,] 0.19158894 0.19158894 0.17239148

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23902713 1.06672634 0.14753966
#>  [2,] 1.32122318 0.24961481 0.46413907
#>  [3,] 0.01594760 0.09652670 0.09652670
#>  [4,] 0.00988970 0.14375033 0.21492086
#>  [5,] 0.09284229 0.08746430 0.09284229
#>  [6,] 0.57167996 0.08169270 0.46296439
#>  [7,] 0.18269134 0.30939213 0.04187959
#>  [8,] 0.16060206 0.16060206 0.16060206
#>  [9,] 0.15778993 0.04933439 0.04986581
#> [10,] 0.36594475 0.09664458 0.21432603
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.54318337 0.9306348 0.3248251
#>  [2,] 0.08566093 0.1522536 0.1522536
#>  [3,] 0.20467034 0.2046703 0.2046703
#>  [4,] 0.93528617 1.7802947 0.2397904
#>  [5,] 0.31846238 0.3875708 0.9321616
#>  [6,] 0.49419418 0.4941942 0.4192292
#>  [7,] 0.21697076 0.2169708 0.2169708
#>  [8,] 0.01354949 0.2787941 0.8441901
#>  [9,] 0.15846083 0.2495337 0.1745919
#> [10,] 0.54293038 0.1265393 0.6700708
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11267235 0.11267235 0.11267235
#>  [2,] 0.08105141 0.08105141 0.08105141
#>  [3,] 0.03180370 0.03180370 0.03180370
#>  [4,] 0.50658453 0.06762796 0.05939682
#>  [5,] 0.75167229 0.58897372 0.05092953
#>  [6,] 0.02649339 0.34440698 0.22351938
#>  [7,] 0.14854082 1.17740835 0.14854082
#>  [8,] 0.06079646 0.04909624 0.60730874
#>  [9,] 0.25894987 1.10268922 0.93244620
#> [10,] 0.12473864 0.12473864 0.12473864
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.15747709 0.05471632 0.005645238
#>  [2,] 0.35053766 0.15091097 0.725961173
#>  [3,] 0.13265770 0.13265770 0.132657703
#>  [4,] 0.39724282 0.42383021 0.523545626
#>  [5,] 0.20963551 0.20963551 0.704880013
#>  [6,] 0.01486156 0.28261388 0.315662627
#>  [7,] 0.44780928 0.45721033 0.457210330
#>  [8,] 3.05324580 0.57446820 0.402010416
#>  [9,] 1.21040762 0.16105898 0.020690324
#> [10,] 0.12764799 0.08414772 1.533500268

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.62115248 0.62115248 0.6211525
#>  [2,] 0.09405069 0.42029887 0.1606409
#>  [3,] 0.24821448 0.24821448 0.2832937
#>  [4,] 0.01282918 0.12723611 0.1901370
#>  [5,] 1.56842167 0.39769003 0.9515295
#>  [6,] 0.16417607 0.16747898 0.1674790
#>  [7,] 0.23425820 0.15490606 0.5252805
#>  [8,] 0.41146268 0.01288499 0.4114627
#>  [9,] 0.29131066 0.20689678 0.1669231
#> [10,] 0.78357995 0.52724811 0.7835800
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.70538117 0.70538117 0.47268877
#>  [2,] 0.81997605 0.70523455 0.78903183
#>  [3,] 1.09144391 1.09144391 0.63699644
#>  [4,] 0.13888095 0.13888095 0.10120671
#>  [5,] 0.02346738 0.08071952 0.08071952
#>  [6,] 0.08244273 0.34055661 0.29796795
#>  [7,] 0.35344548 0.55986151 0.09711282
#>  [8,] 0.22677517 0.79875186 0.58424188
#>  [9,] 0.01687890 0.01687890 0.01687890
#> [10,] 0.05446056 0.05446056 0.05446056
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05586241 0.05586241 0.05586241
#>  [2,] 0.13269422 0.13269422 0.13269422
#>  [3,] 0.38666988 0.19678726 0.15434774
#>  [4,] 0.10913697 0.10237914 0.10913697
#>  [5,] 0.30440133 0.30440133 0.09077274
#>  [6,] 0.09062379 0.63425236 0.09062379
#>  [7,] 0.09355337 0.13989063 0.39344423
#>  [8,] 0.33589328 0.20494290 0.71376942
#>  [9,] 0.11652091 0.48136159 0.48136159
#> [10,] 0.15234220 0.15234220 0.15234220
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06052307 0.06052307 0.06052307
#>  [2,] 0.08808214 0.95240650 0.12768136
#>  [3,] 0.55487976 0.10086925 0.55487976
#>  [4,] 0.07306938 1.65189639 0.47920843
#>  [5,] 1.40338109 0.36237035 0.36237035
#>  [6,] 0.32916081 0.01257557 0.80665426
#>  [7,] 0.10623191 0.10623191 0.10623191
#>  [8,] 0.13509589 0.13509589 0.09940821
#>  [9,] 0.10198758 0.31890305 0.20082337
#> [10,] 0.07734355 0.02348442 0.19432556