Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.7725617 1.7725617 0.3489879
#>  [2,] 4.8569981 0.3096492 5.3387861
#>  [3,] 3.2214153 1.0049234 2.3086784
#>  [4,] 2.5690244 2.1144018 0.2123241
#>  [5,] 2.1064323 1.6720137 2.9115994
#>  [6,] 0.3270134 0.3270134 0.3270134
#>  [7,] 0.9667272 1.9705402 0.3118193
#>  [8,] 1.3519572 0.2630005 0.6050861
#>  [9,] 0.2355654 0.2355654 0.2355654
#> [10,] 4.3012303 0.3441281 3.4847069
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.45757216 2.45757216 2.45757216
#>  [2,] 0.37542027 0.37542027 0.37542027
#>  [3,] 0.90543640 0.90543640 0.90543640
#>  [4,] 0.74106435 0.74106435 0.74106435
#>  [5,] 1.27260404 1.27260404 1.27260404
#>  [6,] 0.20088776 0.20088776 0.20088776
#>  [7,] 0.46442470 0.46442470 0.46442470
#>  [8,] 0.01837397 0.01837397 0.01837397
#>  [9,] 1.54736448 1.54736448 1.54736448
#> [10,] 0.39270284 0.39270284 0.39270284
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.27077484 2.02760733 0.6073217
#>  [2,] 0.45083105 1.22318893 0.7178246
#>  [3,] 2.14364993 0.14085526 0.8850185
#>  [4,] 0.83238280 0.05384548 0.6223529
#>  [5,] 0.15848169 0.28096510 0.5315940
#>  [6,] 0.81034956 0.81658692 0.5850326
#>  [7,] 1.40443168 0.65454174 0.2411255
#>  [8,] 0.03807483 1.21084824 0.1121246
#>  [9,] 1.35548483 0.27893668 0.9711770
#> [10,] 0.88067783 1.25245932 0.0511811

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3172148 0.30114399 1.8309981
#>  [2,] 0.3486134 1.54953889 0.2112021
#>  [3,] 1.3176240 2.77016289 3.6051943
#>  [4,] 0.3302216 0.05211242 0.7632347
#>  [5,] 2.2423457 0.86165009 0.4858221
#>  [6,] 0.8880700 3.01020839 0.6318114
#>  [7,] 0.4290366 1.10075731 5.3995446
#>  [8,] 0.2810855 0.28108546 0.2810855
#>  [9,] 8.5241635 3.12176916 0.2961273
#> [10,] 2.8860039 1.84177738 2.6762163
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.78720520 0.78720520 0.78720520
#>  [2,] 1.31350376 1.31350376 1.31350376
#>  [3,] 2.31779369 2.31779369 2.31779369
#>  [4,] 0.84483149 0.84483149 0.84483149
#>  [5,] 0.13251477 0.13251477 0.13251477
#>  [6,] 0.49823266 0.49823266 0.49823266
#>  [7,] 3.23117258 3.23117258 3.23117258
#>  [8,] 1.04436962 1.04436962 1.04436962
#>  [9,] 1.98531003 1.98531003 1.98531003
#> [10,] 0.02025658 0.02025658 0.02025658
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15823991 1.59528865 0.51310048
#>  [2,] 0.31991925 1.02218993 0.28605936
#>  [3,] 2.58786399 0.13193559 3.59319756
#>  [4,] 0.01317271 0.72448098 1.10355602
#>  [5,] 0.34932419 2.67977589 1.07754911
#>  [6,] 0.17605343 0.87809308 0.61659189
#>  [7,] 0.45816105 1.82543961 0.09647555
#>  [8,] 0.68505332 3.38018290 0.31811562
#>  [9,] 0.19967971 8.00154445 0.57239919
#> [10,] 0.84824799 0.09532244 0.41028269

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.291840146 0.5711645 1.2918401
#>  [2,] 0.904162611 0.9173786 1.4476767
#>  [3,] 0.871365922 2.7856313 2.4575125
#>  [4,] 1.458877804 4.3370221 0.3051157
#>  [5,] 0.447144637 1.2209078 0.1022344
#>  [6,] 3.409272816 1.2766900 0.1075712
#>  [7,] 0.068816536 0.2597100 1.0915955
#>  [8,] 0.580933329 0.5809333 0.5809333
#>  [9,] 2.572562452 3.0302403 3.2263910
#> [10,] 0.007416515 0.1049803 0.1049803
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.62575128 0.62575128 0.62575128
#>  [2,] 1.86136583 1.86136583 1.86136583
#>  [3,] 0.67945720 0.67945720 0.67945720
#>  [4,] 0.20742728 0.20742728 0.20742728
#>  [5,] 0.71376909 0.71376909 0.71376909
#>  [6,] 0.79732359 0.79732359 0.79732359
#>  [7,] 0.08547953 0.08547953 0.08547953
#>  [8,] 0.68016125 0.68016125 0.68016125
#>  [9,] 0.11775159 0.11775159 0.11775159
#> [10,] 1.80497441 1.80497441 1.80497441
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]        [,2]        [,3]
#>  [1,] 1.59625018 0.007994551 0.035586141
#>  [2,] 0.82817495 0.155544703 0.166473185
#>  [3,] 0.03915689 0.975478196 2.659041529
#>  [4,] 0.19854294 0.164782535 0.188520879
#>  [5,] 1.20645910 3.778618743 1.275795439
#>  [6,] 0.95323974 1.278191139 0.388848518
#>  [7,] 0.72665060 0.144461081 0.556209934
#>  [8,] 0.45665945 0.818519275 0.003674251
#>  [9,] 0.28977661 0.696717957 1.267470777
#> [10,] 0.37236177 0.102533637 0.406313242

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.91638869 0.57294261 3.23238183
#>  [2,] 0.76213676 0.71760794 0.76213676
#>  [3,] 0.78381404 0.78381404 0.78381404
#>  [4,] 0.03080257 0.03080257 0.03080257
#>  [5,] 3.63851679 5.28347734 0.09614386
#>  [6,] 0.03209081 0.21433728 0.21433728
#>  [7,] 1.05884899 1.05884899 1.05884899
#>  [8,] 1.00912398 1.14236982 1.14236982
#>  [9,] 3.98581185 0.17620079 1.34073294
#> [10,] 2.58344432 4.12095844 0.63939900
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6737422 1.6737422 1.6737422
#>  [2,] 0.6359276 0.6359276 0.6359276
#>  [3,] 0.1707733 0.1707733 0.1707733
#>  [4,] 3.9396109 3.9396109 3.9396109
#>  [5,] 1.5452602 1.5452602 1.5452602
#>  [6,] 1.1098670 1.1098670 1.1098670
#>  [7,] 0.3589448 0.3589448 0.3589448
#>  [8,] 0.9263850 0.9263850 0.9263850
#>  [9,] 0.3278899 0.3278899 0.3278899
#> [10,] 2.8690727 2.8690727 2.8690727
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 2.22990031 3.13466561 0.8597483
#>  [2,] 0.13045112 0.92010474 3.6781892
#>  [3,] 0.03137693 1.50581036 0.0801723
#>  [4,] 0.10137582 0.49757831 1.2980124
#>  [5,] 2.70815365 0.04741096 0.7978520
#>  [6,] 1.26798318 0.52335286 0.3917928
#>  [7,] 3.07722352 2.20956895 0.8274619
#>  [8,] 0.13825836 0.55047043 2.1391616
#>  [9,] 0.27997617 0.36208375 0.1703221
#> [10,] 0.12370700 0.01487233 0.2536031

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]        [,3]
#>  [1,] 5.00509384 0.42988810 3.850743805
#>  [2,] 0.02045959 0.02045959 0.020459589
#>  [3,] 0.03822064 0.75288802 1.450647074
#>  [4,] 0.71138840 0.08567097 0.474448985
#>  [5,] 2.21012662 1.15477716 2.409157966
#>  [6,] 0.44020242 0.44020242 0.418133831
#>  [7,] 0.24447865 0.16907501 5.184668958
#>  [8,] 1.66762163 0.73124553 0.609847397
#>  [9,] 2.99718569 0.62842599 1.819100045
#> [10,] 2.15136487 0.32026675 0.009823512
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1893788 0.1893788 0.1893788
#>  [2,] 0.3004379 0.3004379 0.3004379
#>  [3,] 0.4612918 0.4612918 0.4612918
#>  [4,] 0.3937685 0.3937685 0.3937685
#>  [5,] 0.2068359 0.2068359 0.2068359
#>  [6,] 0.1289860 0.1289860 0.1289860
#>  [7,] 0.7393526 0.7393526 0.7393526
#>  [8,] 0.1180184 0.1180184 0.1180184
#>  [9,] 1.5789151 1.5789151 1.5789151
#> [10,] 0.1430021 0.1430021 0.1430021
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>              [,1]       [,2]      [,3]
#>  [1,] 1.036554245 1.05741750 0.8123670
#>  [2,] 0.926733079 1.87388229 0.5965419
#>  [3,] 0.784221468 1.28444968 0.2777953
#>  [4,] 0.067856237 1.20990241 2.3462446
#>  [5,] 0.369476505 1.09378389 0.5059729
#>  [6,] 0.002022904 0.08096496 0.3509973
#>  [7,] 0.606075930 0.34164460 0.1047290
#>  [8,] 1.604042503 0.11760526 0.5401050
#>  [9,] 0.014911675 1.55681341 2.1893817
#> [10,] 1.391863504 0.64336148 0.7175016

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.05803543 1.7588177 2.74157640
#>  [2,] 3.41195119 3.4119512 0.95527867
#>  [3,] 3.56154940 3.5615494 1.04062012
#>  [4,] 1.80006767 0.5360961 0.67941402
#>  [5,] 0.95148865 1.4657334 0.05466625
#>  [6,] 2.98341395 0.5850054 2.98341395
#>  [7,] 0.32449995 1.6079530 4.37389662
#>  [8,] 1.12965277 0.4751667 1.25056820
#>  [9,] 0.05874794 0.4507769 0.45077689
#> [10,] 1.87976093 0.3927699 3.78912330
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.43723352 1.43723352 1.43723352
#>  [2,] 0.15955281 0.15955281 0.15955281
#>  [3,] 0.24103454 0.24103454 0.24103454
#>  [4,] 0.49100969 0.49100969 0.49100969
#>  [5,] 0.06191662 0.06191662 0.06191662
#>  [6,] 0.94232740 0.94232740 0.94232740
#>  [7,] 0.68728665 0.68728665 0.68728665
#>  [8,] 0.35900947 0.35900947 0.35900947
#>  [9,] 0.81303660 0.81303660 0.81303660
#> [10,] 0.93626073 0.93626073 0.93626073
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.76532501 0.35388564 0.57027291
#>  [2,] 0.23143614 1.17392985 0.06595194
#>  [3,] 0.71213167 0.98164627 2.41641353
#>  [4,] 0.01971947 0.07756999 1.64293433
#>  [5,] 0.68852331 4.62292104 4.09224395
#>  [6,] 0.12667781 0.10613204 1.61578819
#>  [7,] 1.16748124 1.73972992 1.31580792
#>  [8,] 0.26734483 0.04380373 1.86360820
#>  [9,] 0.56041090 5.76909159 0.15924060
#> [10,] 1.06890682 1.66516226 0.58149296

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0951893 0.5619458 0.62757687
#>  [2,] 3.7345395 0.7584430 3.60901250
#>  [3,] 1.7782128 0.5708708 0.05855111
#>  [4,] 0.6433549 0.1114860 0.64404110
#>  [5,] 0.7100542 6.2364828 2.24566406
#>  [6,] 2.7376106 1.0151913 1.35737340
#>  [7,] 0.8036176 7.2191461 4.42080639
#>  [8,] 0.1954300 0.5331565 0.01097808
#>  [9,] 0.9690205 0.2795276 1.00544195
#> [10,] 3.2258689 1.2129943 1.98632686
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.79556595 0.79556595 0.79556595
#>  [2,] 0.12392929 0.12392929 0.12392929
#>  [3,] 0.75390310 0.75390310 0.75390310
#>  [4,] 0.69680166 0.69680166 0.69680166
#>  [5,] 1.12734677 1.12734677 1.12734677
#>  [6,] 0.06975438 0.06975438 0.06975438
#>  [7,] 6.19385670 6.19385670 6.19385670
#>  [8,] 1.23604831 1.23604831 1.23604831
#>  [9,] 0.45478038 0.45478038 0.45478038
#> [10,] 0.35255640 0.35255640 0.35255640
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5436575 0.08091991 1.1641123
#>  [2,] 0.2414674 0.29627425 1.3397284
#>  [3,] 1.8951760 1.52735275 0.6807864
#>  [4,] 1.4178603 0.99730747 0.1128088
#>  [5,] 2.5101453 0.08453155 1.3541498
#>  [6,] 1.3031500 0.81172000 0.1107906
#>  [7,] 1.2896791 2.62334296 1.1833821
#>  [8,] 0.1509113 1.85910855 1.0008929
#>  [9,] 0.2062760 0.26107817 0.1065028
#> [10,] 0.7279524 0.85580018 0.5773970

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 4.1306880 0.18436068 1.6235498
#>  [2,] 0.1005839 2.04482967 1.2434716
#>  [3,] 2.3757984 2.65294888 2.6529489
#>  [4,] 2.1292950 0.93434424 0.2393713
#>  [5,] 1.0503934 1.73285726 0.3900967
#>  [6,] 1.9657686 3.28926165 1.5593970
#>  [7,] 1.7926869 1.12357638 2.9600992
#>  [8,] 0.8976388 1.57942736 1.5794274
#>  [9,] 0.3092102 0.30921018 0.3092102
#> [10,] 3.4408636 0.09813905 1.1287259
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.92506123 0.92506123 0.92506123
#>  [2,] 0.02015684 0.02015684 0.02015684
#>  [3,] 0.66018489 0.66018489 0.66018489
#>  [4,] 0.67532345 0.67532345 0.67532345
#>  [5,] 0.71258656 0.71258656 0.71258656
#>  [6,] 1.91017911 1.91017911 1.91017911
#>  [7,] 0.02258434 0.02258434 0.02258434
#>  [8,] 0.34719666 0.34719666 0.34719666
#>  [9,] 0.26503092 0.26503092 0.26503092
#> [10,] 1.01644091 1.01644091 1.01644091
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 2.73209183 0.49563983 1.9013468
#>  [2,] 0.53789104 1.13068979 3.3243656
#>  [3,] 2.65061873 0.48003951 0.9540790
#>  [4,] 1.02080429 0.26917665 0.5876110
#>  [5,] 1.41874550 0.08508401 0.6105623
#>  [6,] 0.56714259 0.15222322 0.6220736
#>  [7,] 2.38449865 0.85552438 0.5119407
#>  [8,] 0.66704509 2.46314537 0.7375027
#>  [9,] 0.27921125 0.09470945 0.4686327
#> [10,] 0.07782416 0.13240017 1.0536460

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 2.1398893 3.80869508 1.1015242
#>  [2,] 0.8183330 0.82025355 0.9104314
#>  [3,] 2.7420983 0.88807639 0.9371907
#>  [4,] 1.6075500 0.23572779 3.3856699
#>  [5,] 4.2727300 0.08565972 2.5607454
#>  [6,] 0.1599817 0.81969223 0.2073699
#>  [7,] 1.1558974 0.19112947 0.9851613
#>  [8,] 3.5307967 2.09743503 0.7366324
#>  [9,] 1.2517920 1.25179204 1.2517920
#> [10,] 0.7756119 0.39124914 5.2860675
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 3.33114886 3.33114886 3.33114886
#>  [2,] 0.05870562 0.05870562 0.05870562
#>  [3,] 0.27006799 0.27006799 0.27006799
#>  [4,] 0.44313285 0.44313285 0.44313285
#>  [5,] 0.09662008 0.09662008 0.09662008
#>  [6,] 0.07532850 0.07532850 0.07532850
#>  [7,] 0.50895882 0.50895882 0.50895882
#>  [8,] 2.68106354 2.68106354 2.68106354
#>  [9,] 0.59256953 0.59256953 0.59256953
#> [10,] 3.80901056 3.80901056 3.80901056
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1589746 1.50689981 1.42356991
#>  [2,] 1.3823607 1.10552822 0.65727257
#>  [3,] 1.0749310 0.18836917 0.08027223
#>  [4,] 2.1413090 0.29295342 0.03156319
#>  [5,] 0.4735792 0.10231377 0.58262611
#>  [6,] 2.4740143 0.01561323 1.00908800
#>  [7,] 2.6097546 3.43500828 5.15674599
#>  [8,] 0.9936883 0.06815752 1.04094480
#>  [9,] 0.4975460 1.25287300 0.70667349
#> [10,] 0.3407599 1.16676756 0.03393910

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.85777007 2.7760965 2.2119687
#>  [2,] 0.06399549 0.7920184 0.2254185
#>  [3,] 0.07011944 2.1127007 0.2881419
#>  [4,] 1.37677039 2.0530466 1.2407735
#>  [5,] 0.69752401 0.8211288 1.9671233
#>  [6,] 0.97471664 1.2886175 1.2886175
#>  [7,] 0.36818821 1.0643731 0.1566562
#>  [8,] 2.66756866 4.9566683 1.3446228
#>  [9,] 0.13868539 0.1386854 0.1386854
#> [10,] 1.10673008 0.1448130 4.2218436
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.71197336 0.71197336 0.71197336
#>  [2,] 1.48119927 1.48119927 1.48119927
#>  [3,] 2.71182887 2.71182887 2.71182887
#>  [4,] 0.87476526 0.87476526 0.87476526
#>  [5,] 1.29682345 1.29682345 1.29682345
#>  [6,] 0.26999803 0.26999803 0.26999803
#>  [7,] 0.16919045 0.16919045 0.16919045
#>  [8,] 2.28869246 2.28869246 2.28869246
#>  [9,] 0.16476132 0.16476132 0.16476132
#> [10,] 0.01861553 0.01861553 0.01861553
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42951443 1.14771886 0.99183887
#>  [2,] 2.25815991 0.60258948 1.78430418
#>  [3,] 0.18702337 1.15341195 0.16512474
#>  [4,] 5.16768543 0.61226923 2.99795005
#>  [5,] 4.40092505 0.54966600 1.33785017
#>  [6,] 0.59371120 0.74564443 0.28934556
#>  [7,] 0.07969073 0.13212532 0.05484524
#>  [8,] 0.49133503 0.05041977 1.15428966
#>  [9,] 0.33764166 0.21014264 1.04352413
#> [10,] 1.94417084 0.32150699 0.21072603

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.01913910 1.01913910 0.49537356
#>  [2,] 0.44457830 0.07027043 0.07027043
#>  [3,] 0.81414955 0.11525191 0.11525191
#>  [4,] 0.24804702 0.79543797 0.24804702
#>  [5,] 0.07091842 0.07091842 0.07091842
#>  [6,] 0.47346382 0.59034726 0.89413912
#>  [7,] 0.98750470 1.19407349 0.98750470
#>  [8,] 0.87157986 0.87157986 0.87157986
#>  [9,] 0.02567730 0.02567730 0.64415066
#> [10,] 0.32665879 0.73211753 1.20135224
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.9124473 1.03660237 0.24340219
#>  [2,] 0.6679193 0.06097549 0.20847971
#>  [3,] 1.8141299 0.15795009 0.15795009
#>  [4,] 0.9135317 0.77649322 0.04979401
#>  [5,] 0.3671205 0.36712053 0.48907512
#>  [6,] 0.2205303 0.50361252 0.22053034
#>  [7,] 0.8742246 0.30683289 1.24603781
#>  [8,] 0.1030556 3.83057700 1.21721429
#>  [9,] 1.6412667 1.46462653 1.12982730
#> [10,] 0.3524224 0.71276007 1.05047106
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.849803784 0.78251295 0.84980378
#>  [2,] 2.028885635 1.14671797 0.47659530
#>  [3,] 0.067442346 1.79988097 0.05709454
#>  [4,] 0.003265494 0.02227370 0.76783861
#>  [5,] 0.001538445 0.70551411 0.06825061
#>  [6,] 0.918766206 0.04478287 0.91876621
#>  [7,] 0.770662333 0.77066233 0.14547239
#>  [8,] 0.551580341 1.28629682 0.31770612
#>  [9,] 0.433540314 0.08614348 0.43354031
#> [10,] 0.823845557 0.54126501 0.82384556
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.48693700 0.19077045 0.4869370
#>  [2,] 0.44090064 1.30450580 0.2784783
#>  [3,] 1.27945369 0.35469573 1.2734763
#>  [4,] 0.26655400 0.08620710 0.0862071
#>  [5,] 0.56841756 1.40093183 0.5684176
#>  [6,] 0.09534633 0.09534633 0.1092879
#>  [7,] 0.17122758 0.17122758 0.1712276
#>  [8,] 0.68097484 1.22239104 1.5570211
#>  [9,] 0.11000118 0.11034526 0.9592429
#> [10,] 0.29574666 0.24415752 0.7553897
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23140199 1.01327186 0.23140199
#>  [2,] 0.30928128 0.30928128 0.26612035
#>  [3,] 0.07014849 0.07014849 0.07014849
#>  [4,] 0.04050408 0.81711239 0.05756596
#>  [5,] 0.44632157 0.11042562 0.51647266
#>  [6,] 0.75693524 2.78251513 0.30778394
#>  [7,] 0.23694798 0.92632284 0.88311065
#>  [8,] 0.64196902 1.00797952 0.12606996
#>  [9,] 1.71550925 0.43587684 0.43587684
#> [10,] 3.21795492 1.86640558 1.86640558

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1593551 1.15935508 1.1593551
#>  [2,] 0.6542733 4.98233201 0.9789249
#>  [3,] 1.0393015 0.09030721 4.5833319
#>  [4,] 2.0075592 0.76314866 0.5713595
#>  [5,] 2.0615469 0.63983335 2.0246329
#>  [6,] 5.7536998 0.01776219 0.3347484
#>  [7,] 0.6862565 1.30983721 0.1763705
#>  [8,] 0.7007526 1.75815518 1.7406821
#>  [9,] 0.1779421 1.86681895 1.6754462
#> [10,] 1.1581652 0.07414222 1.9567186
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3545356 0.48235810 0.3219259
#>  [2,] 0.6944307 0.69443072 0.6029825
#>  [3,] 0.9723689 0.84169705 1.1427246
#>  [4,] 6.5980001 1.18058125 0.8755161
#>  [5,] 1.9127160 3.27384724 0.7169779
#>  [6,] 0.5564330 0.08307518 1.3833369
#>  [7,] 3.5049824 2.07903322 2.3066489
#>  [8,] 1.0355879 0.66273989 2.5202043
#>  [9,] 2.5470179 0.64858041 2.5470179
#> [10,] 0.2576128 0.25761279 0.2576128
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3498493 2.5634596 2.5634596
#>  [2,] 0.4225909 1.5390358 1.2160543
#>  [3,] 0.8097956 0.8097956 0.6045572
#>  [4,] 0.1380150 0.1380150 0.1380150
#>  [5,] 2.1525052 3.0134951 2.9020443
#>  [6,] 1.4308909 0.0121866 0.8527833
#>  [7,] 0.2855935 2.6044567 0.5147753
#>  [8,] 0.4581766 1.3796933 1.3796933
#>  [9,] 2.1605814 0.6924454 3.1176256
#> [10,] 0.4516564 7.3694741 0.3938769
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8619254 0.2673203 1.1686120
#>  [2,] 0.3040525 0.3040525 0.3040525
#>  [3,] 5.1081405 5.1081405 3.1673979
#>  [4,] 0.1068848 3.4721635 0.9727282
#>  [5,] 1.7162544 0.7262832 1.8910708
#>  [6,] 0.8284633 0.8327569 0.5206694
#>  [7,] 0.4962251 2.9004631 0.5285891
#>  [8,] 2.4299646 2.1588926 2.0315866
#>  [9,] 2.2614406 3.1883561 2.8257028
#> [10,] 1.0025352 0.0361864 1.0025352
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4296584 0.7561572 0.5108277
#>  [2,] 0.3502344 0.3502344 0.3502344
#>  [3,] 3.6947049 1.0185366 2.3300445
#>  [4,] 3.2224820 3.2224820 3.2224820
#>  [5,] 3.3427142 0.5016909 1.9914819
#>  [6,] 0.3885211 1.6372107 1.6372107
#>  [7,] 0.7053754 2.7368209 1.2988281
#>  [8,] 2.5001835 4.2243869 1.6041726
#>  [9,] 0.2885152 0.2885152 0.2885152
#> [10,] 1.0628845 1.4051652 1.0628845

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23990635 0.23990635 0.23990635
#>  [2,] 2.17333616 2.26837456 2.26837456
#>  [3,] 0.68207155 0.31085099 0.68207155
#>  [4,] 0.75946728 0.59781260 0.08177392
#>  [5,] 0.13918918 1.04600804 0.13918918
#>  [6,] 0.87017278 0.87017278 0.07368272
#>  [7,] 0.35814936 0.35814936 0.14989740
#>  [8,] 0.08189401 0.08189401 0.08189401
#>  [9,] 1.70631540 0.72290328 1.70631540
#> [10,] 0.26657133 0.38749586 0.48602975
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.312587392 0.312587392 0.312587392
#>  [2,] 0.528674441 0.070015025 0.528674441
#>  [3,] 0.005203648 0.005203648 0.005203648
#>  [4,] 0.251394411 0.251394411 0.251394411
#>  [5,] 1.062751594 1.857961050 1.857961050
#>  [6,] 1.755407304 0.917687978 1.755407304
#>  [7,] 0.091178983 0.091178983 0.091178983
#>  [8,] 0.724503822 0.041471271 1.194631029
#>  [9,] 0.276772242 0.153132774 0.276772242
#> [10,] 0.458581289 0.458581289 0.458581289
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13562620 0.16103477 0.91542062
#>  [2,] 0.01095022 0.01095022 0.01095022
#>  [3,] 0.22041409 0.15229415 0.15229415
#>  [4,] 0.39354269 0.52070735 0.52070735
#>  [5,] 0.19980013 0.02658926 0.19980013
#>  [6,] 0.17017385 0.07231098 0.04586982
#>  [7,] 0.24211491 0.08611924 0.34372536
#>  [8,] 0.61482320 0.61482320 0.61482320
#>  [9,] 0.91270497 0.91270497 0.50766508
#> [10,] 0.57503414 0.75576286 0.75576286
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>            [,1]        [,2]      [,3]
#>  [1,] 0.2946655 0.055932664 0.2390520
#>  [2,] 0.1244878 0.124487796 0.1244878
#>  [3,] 0.2251069 0.078592863 1.0203856
#>  [4,] 0.1854830 0.001924526 0.1854830
#>  [5,] 0.5734422 0.161606467 0.5734422
#>  [6,] 0.4414917 2.750684612 2.1372306
#>  [7,] 0.4661390 0.466138999 0.2377137
#>  [8,] 0.1702387 0.076010017 0.2062713
#>  [9,] 1.5441282 0.315779544 0.3157795
#> [10,] 0.2405785 0.240578481 0.2405785
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.636175548 0.636175548 0.636175548
#>  [2,] 1.842554651 1.293305719 0.686116854
#>  [3,] 0.007308011 0.007308011 0.007308011
#>  [4,] 1.136450648 1.212587613 1.058538181
#>  [5,] 0.482168570 0.341612213 0.341612213
#>  [6,] 0.534277900 0.534277900 0.534277900
#>  [7,] 0.632694971 0.867870806 0.950354025
#>  [8,] 0.088650738 0.088650738 0.088650738
#>  [9,] 0.351248097 0.351248097 0.351248097
#> [10,] 0.018034854 0.018034854 0.018034854

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.10643967 0.07791537 0.1510594
#>  [2,] 0.32513842 0.19855599 0.3251384
#>  [3,] 0.40780153 0.47526484 0.4824758
#>  [4,] 0.29426241 0.02752776 0.2395977
#>  [5,] 0.07366912 0.07366912 0.6937426
#>  [6,] 0.61880080 1.91999653 0.1695672
#>  [7,] 0.97267527 0.97267527 0.3591936
#>  [8,] 0.17144572 0.17144572 0.1714457
#>  [9,] 0.08326404 0.08326404 0.2151017
#> [10,] 0.50052855 0.25045683 0.4104181
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>              [,1]        [,2]      [,3]
#>  [1,] 1.735872395 1.735872395 0.1639604
#>  [2,] 0.837923289 0.615172622 0.6151726
#>  [3,] 0.770589942 0.770589942 0.4638441
#>  [4,] 1.086009839 0.658649330 0.8719939
#>  [5,] 0.801193501 0.361082646 0.3610826
#>  [6,] 0.121870830 0.902700533 0.1218708
#>  [7,] 0.166003051 0.166003051 0.1660031
#>  [8,] 0.128103358 0.128103358 1.0397443
#>  [9,] 0.225555172 0.447818694 0.4478187
#> [10,] 0.008477229 0.008477229 0.2629480
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.36935922 0.2741326 1.03825444
#>  [2,] 0.42261857 0.1068574 0.42261857
#>  [3,] 0.08626496 0.6978015 0.65062932
#>  [4,] 0.01778236 1.1392275 0.46862739
#>  [5,] 0.35518504 0.3551850 0.35518504
#>  [6,] 1.15306254 0.1386148 0.71728992
#>  [7,] 1.46611307 0.1078743 0.76610804
#>  [8,] 0.39966578 0.3996658 0.39966578
#>  [9,] 0.53601293 0.5360129 0.44541894
#> [10,] 0.07908256 0.1269414 0.07908256
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.164122281 0.16412228 0.16412228
#>  [2,] 0.046609811 0.08725548 0.08725548
#>  [3,] 1.046839579 0.55094119 0.19525790
#>  [4,] 0.163700930 0.16370093 0.16370093
#>  [5,] 0.136518237 0.13651824 0.13651824
#>  [6,] 0.442107852 0.26925546 0.26925546
#>  [7,] 0.379806321 0.37980632 0.37980632
#>  [8,] 0.002335132 0.37779911 0.05993086
#>  [9,] 0.300165542 0.01510828 0.04990528
#> [10,] 0.719939897 0.32581841 0.91923140

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 1.09643896 0.1354608 0.12176191
#>  [2,] 0.13904074 0.4898358 1.12427860
#>  [3,] 0.17044154 0.3181934 0.31819338
#>  [4,] 0.74220882 0.8858604 0.15002623
#>  [5,] 0.24630870 0.2463087 0.24630870
#>  [6,] 0.01170361 0.4722899 0.01170361
#>  [7,] 0.33955422 0.4467151 0.45092745
#>  [8,] 1.36791312 0.3729185 0.15311389
#>  [9,] 0.51463681 0.5146368 0.29638979
#> [10,] 0.32402384 0.6583245 0.06850648
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 1.106408600 0.20321489 0.10175350
#>  [2,] 0.471506857 0.02552429 0.38225768
#>  [3,] 0.140256734 0.01413092 0.14025673
#>  [4,] 0.780203004 0.05404600 0.63542079
#>  [5,] 0.486247328 0.57414790 0.18819885
#>  [6,] 0.001937743 0.01893860 0.01108268
#>  [7,] 0.482340539 0.31397013 0.65135713
#>  [8,] 0.102932258 0.33426577 2.16956220
#>  [9,] 0.013014537 0.02390531 0.01301454
#> [10,] 0.515731030 1.19046646 0.66895965
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2936085 0.41993901 0.08736138
#>  [2,] 0.4861283 0.04647566 0.19416985
#>  [3,] 0.6627537 0.17687651 0.97107930
#>  [4,] 0.9695931 0.73483179 0.71210359
#>  [5,] 0.5530019 0.10609530 0.60133709
#>  [6,] 0.2009605 0.46605972 0.11154877
#>  [7,] 0.1926112 0.65524330 0.68581340
#>  [8,] 0.3007203 1.20697086 0.62085444
#>  [9,] 0.1399368 0.10338285 0.51813342
#> [10,] 1.0529290 0.21102831 0.06747717
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07373247 0.07373247 0.07373247
#>  [2,] 0.12187108 0.67764689 0.14897652
#>  [3,] 0.75265917 0.75265917 0.75265917
#>  [4,] 0.06329602 0.08270144 0.08270144
#>  [5,] 0.03486996 0.19017492 0.19017492
#>  [6,] 0.46010035 0.57973335 0.57973335
#>  [7,] 0.30628947 0.30628947 0.19880580
#>  [8,] 0.48092472 0.08641092 0.60262712
#>  [9,] 0.29588194 0.05248484 0.05248484
#> [10,] 0.78047702 0.21981683 0.15079569

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06640257 0.14470658 0.14470658
#>  [2,] 1.18843449 1.18843449 1.18843449
#>  [3,] 0.02089969 0.02089969 0.02089969
#>  [4,] 0.54386939 0.58176478 0.01610504
#>  [5,] 0.04424416 0.04424416 0.04424416
#>  [6,] 0.45763283 0.12884300 0.31802823
#>  [7,] 0.25216841 0.58188950 0.11903749
#>  [8,] 0.90662562 0.48300280 0.65404954
#>  [9,] 0.04943069 0.04943069 0.04943069
#> [10,] 0.08595784 0.08595784 0.08595784
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.386858973 0.386858973 0.263925541
#>  [2,] 0.424366931 0.734678337 0.424366931
#>  [3,] 0.059529102 0.028587420 0.059529102
#>  [4,] 0.399659745 0.399659745 0.529500510
#>  [5,] 0.138020867 0.239999881 0.138020867
#>  [6,] 0.091386908 0.064960504 0.064960504
#>  [7,] 0.008708451 0.008708451 0.008708451
#>  [8,] 0.237972955 0.237972955 0.237972955
#>  [9,] 0.563748650 0.031761741 0.077576899
#> [10,] 0.979126116 0.979126116 0.662487510
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06354739 0.06354739 0.06354739
#>  [2,] 0.10932033 0.12941887 0.20639941
#>  [3,] 0.07567974 0.29199538 0.08297046
#>  [4,] 0.34886077 0.26368841 0.03870335
#>  [5,] 0.39313046 0.16079339 0.08116986
#>  [6,] 0.11266629 0.11266629 0.11266629
#>  [7,] 0.40375743 0.28043395 0.46010281
#>  [8,] 0.02641080 0.02641080 0.37114834
#>  [9,] 0.04736988 0.08447622 0.08447622
#> [10,] 0.14095721 0.43280878 0.42217392
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>              [,1]       [,2]      [,3]
#>  [1,] 0.589458166 0.48338202 0.2957718
#>  [2,] 0.509884418 0.91034292 0.9741504
#>  [3,] 0.813624649 0.83437954 0.2516200
#>  [4,] 0.102624682 0.87953946 0.3592048
#>  [5,] 0.086115653 0.20003048 0.2000305
#>  [6,] 0.028095007 0.26151609 0.1002496
#>  [7,] 0.603375162 1.19351579 1.7735007
#>  [8,] 0.612538173 0.22210251 0.1061436
#>  [9,] 0.206069058 0.20606906 0.2060691
#> [10,] 0.005736257 0.09901581 0.4151770