Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5525224 2.0309253 0.94395236
#>  [2,] 0.9628435 1.3731860 1.96710205
#>  [3,] 2.0928171 0.0662978 2.09281707
#>  [4,] 1.3222186 5.1172234 2.39703846
#>  [5,] 0.9086680 0.1035084 0.35569700
#>  [6,] 2.1903163 0.5483466 0.07154674
#>  [7,] 1.6501758 1.6501758 1.65017579
#>  [8,] 2.4256337 0.8519562 0.58747515
#>  [9,] 2.4080932 2.0015202 1.31979314
#> [10,] 1.3260362 0.5293785 1.95324035
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58911330 0.58911330 0.58911330
#>  [2,] 0.06782126 0.06782126 0.06782126
#>  [3,] 0.97351926 0.97351926 0.97351926
#>  [4,] 0.20053940 0.20053940 0.20053940
#>  [5,] 0.18822173 0.18822173 0.18822173
#>  [6,] 0.52009526 0.52009526 0.52009526
#>  [7,] 0.15594268 0.15594268 0.15594268
#>  [8,] 4.40238985 4.40238985 4.40238985
#>  [9,] 1.98239985 1.98239985 1.98239985
#> [10,] 0.59677912 0.59677912 0.59677912
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24514165 0.30110366 0.34065422
#>  [2,] 0.96532095 0.08700110 0.08594567
#>  [3,] 0.07219788 4.41478770 2.51688732
#>  [4,] 0.79134635 0.32404707 0.89168184
#>  [5,] 0.14130785 1.44024695 1.43333245
#>  [6,] 1.34752102 0.01305032 4.39015775
#>  [7,] 0.24585885 0.68669105 0.24777837
#>  [8,] 0.82772400 3.80142820 2.47305974
#>  [9,] 0.24296624 0.74868788 1.50240016
#> [10,] 0.28058398 0.10651861 0.47569647

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 4.34600049 6.2256362 4.2672584
#>  [2,] 1.25704019 0.2370510 1.9306843
#>  [3,] 0.05428641 0.1079416 1.0383698
#>  [4,] 0.94632537 1.8833043 6.4143112
#>  [5,] 4.88859811 5.4839528 0.8971540
#>  [6,] 0.32594350 0.7436500 0.7436500
#>  [7,] 0.14557698 1.0560755 0.7296734
#>  [8,] 0.11772119 0.9591946 4.0109810
#>  [9,] 0.21897303 0.2189730 0.2189730
#> [10,] 0.03012881 0.8245898 0.7666109
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.45856150 0.45856150 0.45856150
#>  [2,] 0.19626710 0.19626710 0.19626710
#>  [3,] 0.24230091 0.24230091 0.24230091
#>  [4,] 0.56654877 0.56654877 0.56654877
#>  [5,] 0.09558094 0.09558094 0.09558094
#>  [6,] 0.38052925 0.38052925 0.38052925
#>  [7,] 0.15382180 0.15382180 0.15382180
#>  [8,] 0.48078840 0.48078840 0.48078840
#>  [9,] 1.22003746 1.22003746 1.22003746
#> [10,] 0.91681451 0.91681451 0.91681451
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]      [,2]        [,3]
#>  [1,] 1.28609527 1.4786733 1.796487486
#>  [2,] 0.23814784 1.7183553 0.017846118
#>  [3,] 2.68066899 2.4075496 0.010969016
#>  [4,] 0.29924727 1.4157254 0.849166740
#>  [5,] 0.89371856 2.7996260 0.003198337
#>  [6,] 0.01514795 0.7410836 0.490974568
#>  [7,] 4.36766617 1.5381355 1.108582964
#>  [8,] 0.01401094 4.2767802 0.042743984
#>  [9,] 0.88594038 0.7326800 1.137830038
#> [10,] 0.06446883 0.2259116 2.467385074

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28553438 2.77659487 2.77659487
#>  [2,] 0.42146275 0.54085683 0.19814686
#>  [3,] 1.79883682 1.37740914 4.65440902
#>  [4,] 1.39612137 0.28679061 0.06534782
#>  [5,] 1.17443291 1.92603470 7.14878155
#>  [6,] 1.44991769 0.26673017 1.31744808
#>  [7,] 0.04460488 0.04460488 0.04460488
#>  [8,] 0.94748703 2.14770933 0.15847582
#>  [9,] 1.08511160 0.77162296 1.08511160
#> [10,] 1.22573117 2.20081178 0.26011064
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4956926 0.4956926 0.4956926
#>  [2,] 0.7484528 0.7484528 0.7484528
#>  [3,] 1.6584236 1.6584236 1.6584236
#>  [4,] 0.1941145 0.1941145 0.1941145
#>  [5,] 0.4600432 0.4600432 0.4600432
#>  [6,] 0.3515635 0.3515635 0.3515635
#>  [7,] 1.0316841 1.0316841 1.0316841
#>  [8,] 1.6005182 1.6005182 1.6005182
#>  [9,] 0.0221230 0.0221230 0.0221230
#> [10,] 1.3787273 1.3787273 1.3787273
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.99865294 0.92797506 0.7015196
#>  [2,] 2.16030200 0.04596927 0.3904624
#>  [3,] 0.01141935 0.27965627 0.2864670
#>  [4,] 1.23201180 0.01532533 0.3526906
#>  [5,] 1.96318489 0.42790725 0.1318304
#>  [6,] 1.42485032 0.28464475 3.3748810
#>  [7,] 0.17814178 0.26863002 0.8121733
#>  [8,] 0.41091868 1.34893782 1.1264089
#>  [9,] 0.76761386 2.63979354 1.4834256
#> [10,] 0.09710454 2.85498308 0.1373931

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7886829 0.10039037 6.2240329
#>  [2,] 1.1140961 0.44664957 3.9275444
#>  [3,] 0.7184226 2.77269153 1.2560785
#>  [4,] 0.4266821 1.79723011 1.2563577
#>  [5,] 0.2761775 0.28085543 1.9951189
#>  [6,] 0.5826301 0.06352307 0.7014642
#>  [7,] 0.8778260 1.77811044 0.5611990
#>  [8,] 0.9832058 0.27701293 0.5325825
#>  [9,] 1.9217732 2.64037364 1.3484981
#> [10,] 3.1582002 2.74285275 3.6206869
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17219710 0.17219710 0.17219710
#>  [2,] 0.57315562 0.57315562 0.57315562
#>  [3,] 0.64353046 0.64353046 0.64353046
#>  [4,] 0.26599056 0.26599056 0.26599056
#>  [5,] 0.80177687 0.80177687 0.80177687
#>  [6,] 1.03588833 1.03588833 1.03588833
#>  [7,] 0.04261726 0.04261726 0.04261726
#>  [8,] 0.85421878 0.85421878 0.85421878
#>  [9,] 1.39162944 1.39162944 1.39162944
#> [10,] 0.22104089 0.22104089 0.22104089
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9450575 0.8471456 0.17231362
#>  [2,] 0.3652956 0.7814587 0.42056546
#>  [3,] 2.3930921 0.5344061 0.05212262
#>  [4,] 0.4446743 1.7950889 0.72796521
#>  [5,] 2.4115729 0.9441620 0.73072319
#>  [6,] 0.0925798 0.2776823 0.56213514
#>  [7,] 0.3236162 0.2953690 0.16555433
#>  [8,] 1.5344889 0.1378066 0.43016599
#>  [9,] 0.2334521 0.8217662 0.16727414
#> [10,] 1.2383016 0.1672936 0.54531855

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]        [,2]      [,3]
#>  [1,] 0.9708164 3.476287374 0.3365545
#>  [2,] 0.2231263 0.223126344 0.2231263
#>  [3,] 1.4941579 0.001432299 1.4699446
#>  [4,] 0.0104679 0.305786628 0.4026560
#>  [5,] 0.6899361 0.689936060 0.6899361
#>  [6,] 1.3784010 4.458745296 1.3969507
#>  [7,] 0.8197112 1.527795678 1.5277957
#>  [8,] 3.6256263 3.625626306 1.3566703
#>  [9,] 0.6214359 1.081507858 0.4047685
#> [10,] 5.5878204 1.006457311 0.7884630
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 4.5945439 4.5945439 4.5945439
#>  [2,] 2.6443335 2.6443335 2.6443335
#>  [3,] 0.6838708 0.6838708 0.6838708
#>  [4,] 0.1782497 0.1782497 0.1782497
#>  [5,] 0.4299741 0.4299741 0.4299741
#>  [6,] 0.5746054 0.5746054 0.5746054
#>  [7,] 0.4182343 0.4182343 0.4182343
#>  [8,] 0.3702632 0.3702632 0.3702632
#>  [9,] 1.0171231 1.0171231 1.0171231
#> [10,] 0.0783934 0.0783934 0.0783934
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1799596 0.2660179 1.3274753
#>  [2,] 2.5806476 0.6008889 0.5881075
#>  [3,] 0.3163422 2.6707513 0.8213255
#>  [4,] 1.2243768 0.5896652 0.7012472
#>  [5,] 1.3970915 0.6495508 0.0862569
#>  [6,] 0.9646298 0.5837246 0.4469993
#>  [7,] 0.5264979 0.5100236 0.2066579
#>  [8,] 0.1487419 0.1214160 0.6584693
#>  [9,] 0.4127628 0.2034606 1.4960419
#> [10,] 0.2752409 0.4654032 0.3956347

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.1182211 3.2051905 1.02345624
#>  [2,] 0.5545124 0.4904844 0.03713667
#>  [3,] 1.3961394 0.1179511 0.98907923
#>  [4,] 3.0768900 0.8883559 0.44709725
#>  [5,] 0.9245957 0.3124305 2.87702746
#>  [6,] 1.6690560 1.3852656 0.16487028
#>  [7,] 4.1900837 0.4685891 3.10284006
#>  [8,] 0.5765473 0.1944592 0.84051036
#>  [9,] 4.1793191 1.7219910 5.10833667
#> [10,] 4.0242429 0.4592767 3.96385063
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2033147 1.2033147 1.2033147
#>  [2,] 0.3541392 0.3541392 0.3541392
#>  [3,] 0.7685574 0.7685574 0.7685574
#>  [4,] 0.8550588 0.8550588 0.8550588
#>  [5,] 2.3700787 2.3700787 2.3700787
#>  [6,] 0.8599388 0.8599388 0.8599388
#>  [7,] 0.3072986 0.3072986 0.3072986
#>  [8,] 0.8510346 0.8510346 0.8510346
#>  [9,] 0.2916053 0.2916053 0.2916053
#> [10,] 0.3464249 0.3464249 0.3464249
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.9170666 0.1862250 0.30575441
#>  [2,] 0.6905355 0.7627213 0.08916273
#>  [3,] 1.3357285 0.2853883 0.49936811
#>  [4,] 1.3780058 1.8441195 1.35932652
#>  [5,] 1.4481006 0.2484987 0.56050553
#>  [6,] 2.6363803 2.2521483 0.47961902
#>  [7,] 1.1301461 0.1481830 0.33364038
#>  [8,] 0.4586446 4.8341647 1.81779600
#>  [9,] 1.1671344 0.8366196 0.35670157
#> [10,] 1.6616973 0.7862882 1.53607616

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM")
#>              [,1]       [,2]       [,3]
#>  [1,]  0.09871644 0.09871644 0.09871644
#>  [2,]  0.40393934 0.40393934 0.40393934
#>  [3,]  0.04778256 0.04778256 0.04778256
#>  [4,]  3.63858670 0.21687697 0.13799907
#>  [5,]  2.66406040 0.52705397 0.28753366
#>  [6,]  5.32021111 2.57956181 0.51423348
#>  [7,] 10.22734455 4.73986291 0.90929555
#>  [8,]  3.13440786 0.69267195 1.97892400
#>  [9,]  0.42150010 2.15162578 1.06272524
#> [10,]  2.95713476 0.67094320 1.41545596
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.02798827 2.02798827 2.02798827
#>  [2,] 2.25197430 2.25197430 2.25197430
#>  [3,] 0.08549176 0.08549176 0.08549176
#>  [4,] 2.63104293 2.63104293 2.63104293
#>  [5,] 2.03658411 2.03658411 2.03658411
#>  [6,] 0.39559864 0.39559864 0.39559864
#>  [7,] 0.36073830 0.36073830 0.36073830
#>  [8,] 2.28842447 2.28842447 2.28842447
#>  [9,] 1.20784716 1.20784716 1.20784716
#> [10,] 0.03132677 0.03132677 0.03132677
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58833779 0.67815751 1.00793067
#>  [2,] 0.20597214 0.45254598 0.05783827
#>  [3,] 0.63153219 0.06051516 0.28119252
#>  [4,] 0.58442786 0.12455526 0.68887838
#>  [5,] 0.04176431 0.52467326 0.07733960
#>  [6,] 1.87658575 0.83027632 1.02038038
#>  [7,] 0.99272742 1.18920447 0.51761428
#>  [8,] 0.49119089 0.17292349 0.33279914
#>  [9,] 0.37327825 0.70101789 0.50188335
#> [10,] 0.74280173 0.47766705 1.96710118

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3627642 0.01567891 0.36276421
#>  [2,] 0.1543068 0.15430677 0.15430677
#>  [3,] 0.3588149 0.26704007 0.36614645
#>  [4,] 0.6490741 0.64907411 0.64907411
#>  [5,] 1.5646615 2.12658660 1.07081542
#>  [6,] 1.9410338 4.13211809 2.77502411
#>  [7,] 1.1618406 0.01994298 0.15369199
#>  [8,] 1.2644624 0.75145089 0.08017794
#>  [9,] 1.7246346 0.72198422 0.69823896
#> [10,] 1.4555419 2.27534946 4.28256705
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02505883 0.02505883 0.02505883
#>  [2,] 0.44857127 0.44857127 0.44857127
#>  [3,] 1.43444806 1.43444806 1.43444806
#>  [4,] 1.45812800 1.45812800 1.45812800
#>  [5,] 0.41055318 0.41055318 0.41055318
#>  [6,] 2.08562618 2.08562618 2.08562618
#>  [7,] 0.16283916 0.16283916 0.16283916
#>  [8,] 0.34218791 0.34218791 0.34218791
#>  [9,] 1.99557659 1.99557659 1.99557659
#> [10,] 0.70537167 0.70537167 0.70537167
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.76291852 0.8804227 0.43331221
#>  [2,] 0.42913492 0.6770080 1.64844279
#>  [3,] 0.30264450 0.1516588 0.46124874
#>  [4,] 1.11649158 0.2848588 0.08600713
#>  [5,] 0.71440715 0.5807826 0.01947536
#>  [6,] 0.39442611 0.2400903 0.09536000
#>  [7,] 0.97101135 2.8691640 0.31337875
#>  [8,] 0.05109446 0.7698387 3.83329489
#>  [9,] 0.34386841 0.3607169 0.46408787
#> [10,] 2.85912899 0.5596750 0.59944716

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.6239314 0.5755944 0.12885100
#>  [2,] 0.9012774 1.0292630 0.34188870
#>  [3,] 2.9445679 3.6014559 1.17797368
#>  [4,] 5.0508663 0.7347506 0.71081050
#>  [5,] 8.6757647 2.4118839 5.55808959
#>  [6,] 5.3047642 2.5640029 1.23018034
#>  [7,] 0.6813424 1.2724994 2.25300456
#>  [8,] 1.4828233 3.2087147 1.15283958
#>  [9,] 2.6385427 0.4130287 0.06194466
#> [10,] 1.0544066 2.9927862 0.89639347
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.81534240 0.81534240 0.81534240
#>  [2,] 0.69188495 0.69188495 0.69188495
#>  [3,] 0.83110880 0.83110880 0.83110880
#>  [4,] 1.15794656 1.15794656 1.15794656
#>  [5,] 0.22545018 0.22545018 0.22545018
#>  [6,] 0.07724506 0.07724506 0.07724506
#>  [7,] 2.72112598 2.72112598 2.72112598
#>  [8,] 0.60244028 0.60244028 0.60244028
#>  [9,] 0.81483874 0.81483874 0.81483874
#> [10,] 0.49675689 0.49675689 0.49675689
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.31986039 0.01554837 1.28327917
#>  [2,] 0.24164913 0.96112454 0.95618890
#>  [3,] 1.26696467 0.65460576 0.54265562
#>  [4,] 0.02923399 0.83711918 0.10305653
#>  [5,] 0.64320568 0.59501926 0.03205027
#>  [6,] 0.42387147 1.53407408 1.50616804
#>  [7,] 0.87352675 1.11975693 0.17390655
#>  [8,] 1.54672338 0.19577491 1.08141766
#>  [9,] 0.15839333 3.93954809 1.51191581
#> [10,] 2.01050177 4.43744151 0.22238790

rpextmo(
    10, 3, a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.0515930 3.0449825 0.05790076
#>  [2,] 0.5898750 2.3825590 0.48744161
#>  [3,] 0.5369382 0.2873899 0.37735037
#>  [4,] 3.6871237 0.6448120 1.13289816
#>  [5,] 1.3327406 1.2172463 0.35471182
#>  [6,] 0.3703686 2.5277794 1.50925796
#>  [7,] 0.0626223 3.0554913 2.00566162
#>  [8,] 2.0809391 0.5702249 0.01325945
#>  [9,] 0.1098729 0.1098729 0.10987294
#> [10,] 1.3580668 1.2669110 0.32570262
## comonotone
rpextmo(
    10, 3, a = 1,
    family = "Armageddon",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.05002856 2.05002856 2.05002856
#>  [2,] 0.33272441 0.33272441 0.33272441
#>  [3,] 0.09869857 0.09869857 0.09869857
#>  [4,] 3.68144486 3.68144486 3.68144486
#>  [5,] 1.78786690 1.78786690 1.78786690
#>  [6,] 0.04826613 0.04826613 0.04826613
#>  [7,] 1.14462001 1.14462001 1.14462001
#>  [8,] 1.88651949 1.88651949 1.88651949
#>  [9,] 0.88127567 0.88127567 0.88127567
#> [10,] 0.34288894 0.34288894 0.34288894
## independence
rpextmo(
    10, 3, b = 1,
    family = "Armageddon",
    method = "AM")
#>              [,1]       [,2]        [,3]
#>  [1,] 3.694496266 0.84332620 1.911383417
#>  [2,] 1.682903326 1.48738467 1.856671945
#>  [3,] 0.428890541 0.28558136 0.358849067
#>  [4,] 0.386760562 0.75667125 0.054948427
#>  [5,] 0.976097192 0.16357861 0.008852583
#>  [6,] 0.597821888 0.05374092 0.518560031
#>  [7,] 0.252584621 1.77907372 2.929701643
#>  [8,] 0.946114630 0.61963376 0.342250973
#>  [9,] 1.241367808 1.39855391 0.952568494
#> [10,] 0.002079787 0.33455537 0.583812178

## Poisson
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson")
#>              [,1]        [,2]        [,3]
#>  [1,] 1.686562948 1.686562948 0.469285766
#>  [2,] 0.007699443 0.007699443 0.007699443
#>  [3,] 0.502063327 0.967236381 0.310629755
#>  [4,] 0.464946353 0.330197707 0.493887204
#>  [5,] 5.207918955 0.161889585 1.238888030
#>  [6,] 0.086396992 0.086396992 0.086396992
#>  [7,] 0.549586496 0.549586496 0.541216032
#>  [8,] 0.351179936 0.153040748 0.603314645
#>  [9,] 0.057845158 0.664952275 1.957411201
#> [10,] 0.225224696 0.519530190 0.519530190
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.075516883 0.075516883 1.135272740
#>  [2,] 0.365947793 0.992504986 0.435085648
#>  [3,] 0.214402847 0.737186285 0.737186285
#>  [4,] 1.079525033 0.844297952 1.042942506
#>  [5,] 0.099506245 0.099506245 0.099506245
#>  [6,] 0.303028893 0.442314013 1.254563369
#>  [7,] 0.007893885 0.007893885 0.007893885
#>  [8,] 1.410952697 0.269031461 1.410952697
#>  [9,] 2.402355911 0.547300101 2.037001189
#> [10,] 0.028973953 0.093609234 0.028973953
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.49366133 0.42008109 0.4936613
#>  [2,] 1.09047688 0.87618638 0.3348647
#>  [3,] 0.47383168 0.23225137 0.0695148
#>  [4,] 0.35113195 0.35113195 0.3511320
#>  [5,] 0.28438597 0.70245221 1.2536411
#>  [6,] 0.16292327 2.10578617 0.2996036
#>  [7,] 0.04125473 0.04125473 0.1508461
#>  [8,] 1.36843334 0.21811250 0.2181125
#>  [9,] 0.04940433 4.71749495 0.1690047
#> [10,] 1.25683797 2.28335175 0.5681197
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM")
#>              [,1]      [,2]       [,3]
#>  [1,] 1.402320839 1.7260474 0.64573134
#>  [2,] 0.000584125 0.6339591 1.43414916
#>  [3,] 1.199480605 0.5620675 2.23579149
#>  [4,] 0.475510787 0.6629819 0.78687162
#>  [5,] 1.221742071 0.5627396 0.28531384
#>  [6,] 1.360952016 0.7205546 0.75747465
#>  [7,] 0.947795975 0.9477960 0.36542912
#>  [8,] 0.479714524 0.4797145 0.01032863
#>  [9,] 0.169560527 0.4636608 0.46366081
#> [10,] 1.842109285 0.6210130 1.32610734
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.94869515 0.43983379 0.62911045
#>  [2,] 1.17397362 1.24610900 1.10335401
#>  [3,] 0.13482189 0.13482189 0.13482189
#>  [4,] 1.29519790 0.59644033 0.16392062
#>  [5,] 1.68303478 0.03843525 1.68303478
#>  [6,] 0.04093756 0.04093756 0.04093756
#>  [7,] 1.07225787 0.65340497 1.48731127
#>  [8,] 0.86081778 0.86081778 0.47257213
#>  [9,] 0.42685901 0.31613995 0.47460077
#> [10,] 0.40763700 0.18081572 0.62857911

## Pareto
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.3155198 1.8228961 2.8120638
#>  [2,] 1.0100463 1.0100463 0.2893006
#>  [3,] 1.0037553 0.2579637 0.4445790
#>  [4,] 2.0784586 1.1543979 5.1475415
#>  [5,] 0.7284968 0.6667701 0.6555808
#>  [6,] 4.3617938 0.1056044 1.5309514
#>  [7,] 0.6621417 0.7860994 0.1676226
#>  [8,] 0.5926482 2.7235220 2.0208920
#>  [9,] 1.8085316 1.4660211 1.1906776
#> [10,] 0.2610772 0.2279662 0.2610772
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6537234 0.66652387 0.6665239
#>  [2,] 0.6927589 2.91902285 0.9768548
#>  [3,] 0.5737397 2.69338697 1.2407873
#>  [4,] 3.4654088 0.15046057 0.7046847
#>  [5,] 0.1938520 0.05694506 0.1938520
#>  [6,] 1.2633923 1.49717170 1.2949513
#>  [7,] 0.3245077 0.67789658 0.6503922
#>  [8,] 0.2530021 0.13372526 0.6474870
#>  [9,] 1.1171034 3.14053786 4.3529827
#> [10,] 0.4208983 1.90568042 6.7031590
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3878503 0.38785030 0.1985740
#>  [2,] 2.4208984 2.42089845 2.4208984
#>  [3,] 0.8040684 3.39371712 4.2186375
#>  [4,] 2.1085236 0.51879059 0.9153864
#>  [5,] 0.8556140 2.57152458 0.3195549
#>  [6,] 1.2311963 0.08489683 0.3464098
#>  [7,] 0.2771137 1.16810877 0.8873967
#>  [8,] 0.1057892 1.16320523 0.6230343
#>  [9,] 0.6563230 1.50337243 0.4770893
#> [10,] 0.7737182 1.40386893 4.8724709
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.86763797 0.8676380 0.2341682
#>  [2,] 0.47548554 0.4332721 1.1256668
#>  [3,] 3.03353260 1.3710837 0.8677638
#>  [4,] 0.06031805 3.9520146 0.1328604
#>  [5,] 4.05443481 2.9300148 0.2220956
#>  [6,] 0.61453398 0.9783978 0.1840400
#>  [7,] 0.14415941 0.4760263 0.8173130
#>  [8,] 0.94104056 0.1950342 1.6835213
#>  [9,] 0.96735276 7.6780041 8.4395985
#> [10,] 0.07504019 1.0685296 1.0685296
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 3.2842897 0.17171982 2.5569884
#>  [2,] 0.5525366 1.48667042 4.5075036
#>  [3,] 1.3436438 0.19316666 4.2875618
#>  [4,] 0.8670047 2.29661503 0.3284279
#>  [5,] 0.7394893 0.35935500 1.5518204
#>  [6,] 2.4734123 2.38523194 1.3211528
#>  [7,] 0.1445007 1.95855317 3.3815511
#>  [8,] 4.8393268 1.42559915 2.4429661
#>  [9,] 1.3696772 0.08733634 4.5614245
#> [10,] 0.8111990 2.01917478 1.4442601

## Exponential
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39303332 0.11628792 0.39303332
#>  [2,] 0.84678987 0.54547639 0.84678987
#>  [3,] 0.32220947 0.32220947 0.03070605
#>  [4,] 0.22189146 0.27492314 0.22189146
#>  [5,] 0.15657555 0.15657555 0.15657555
#>  [6,] 0.74353659 0.72815140 0.78915756
#>  [7,] 0.51648332 0.39187889 0.29350751
#>  [8,] 0.21285421 0.21285421 0.21285421
#>  [9,] 0.04724256 0.04724256 0.04724256
#> [10,] 0.03374642 0.26594573 0.03374642
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5330528 0.53305278 0.53305278
#>  [2,] 0.8301840 1.42103979 0.53562133
#>  [3,] 0.6288638 1.62597058 2.13970212
#>  [4,] 0.4651759 0.12795304 0.46517593
#>  [5,] 0.2597880 0.08357743 0.25978800
#>  [6,] 0.4390362 0.34138389 0.03869555
#>  [7,] 0.2037240 0.20372403 0.20372403
#>  [8,] 0.6637633 1.01895145 1.01895145
#>  [9,] 0.2611602 0.26116020 0.26116020
#> [10,] 0.4093761 0.03682391 0.40937614
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24156065 0.37064744 0.37064744
#>  [2,] 0.05681126 0.05681126 0.05681126
#>  [3,] 0.12381380 0.12381380 0.12381380
#>  [4,] 0.18889779 0.37683260 0.18889779
#>  [5,] 0.35706151 0.03107410 0.35706151
#>  [6,] 0.46058086 0.02790736 0.02790736
#>  [7,] 0.13991649 0.13991649 0.13991649
#>  [8,] 0.11684850 1.09236338 1.09236338
#>  [9,] 0.53853374 0.53853374 0.37644508
#> [10,] 0.53383289 0.37281254 1.06886577
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.85509927 0.31485648 0.85509927
#>  [2,] 1.76735038 2.07793479 2.73855033
#>  [3,] 0.02710334 0.86923205 1.02064492
#>  [4,] 0.12834902 0.12834902 0.12834902
#>  [5,] 2.06077605 0.96867038 1.54001658
#>  [6,] 0.49400061 0.49400061 0.41521400
#>  [7,] 0.06041328 0.06041328 0.06041328
#>  [8,] 1.07485193 1.07485193 1.00999028
#>  [9,] 0.42000861 0.92274642 0.92274642
#> [10,] 0.01672634 0.68370629 0.20551614
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21637026 0.21637026 0.21637026
#>  [2,] 0.45126205 0.45126205 0.45126205
#>  [3,] 0.50127814 0.50127814 0.06242635
#>  [4,] 0.62162074 0.02572790 0.69621966
#>  [5,] 0.30857389 0.30857389 0.05460916
#>  [6,] 0.74615513 0.74615513 0.74615513
#>  [7,] 0.17255762 0.17255762 0.17255762
#>  [8,] 0.01479089 0.01479089 0.01479089
#>  [9,] 0.66481140 0.66481140 0.29887953
#> [10,] 0.39862106 0.90549549 0.12931229

## Alpha-Stable
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.759879619 0.396968957 1.05525866
#>  [2,] 0.528786025 0.551298560 0.52878603
#>  [3,] 0.032091597 0.032091597 0.03209160
#>  [4,] 0.095114761 0.095114761 0.09511476
#>  [5,] 0.198816676 0.070700499 0.81233659
#>  [6,] 0.839661344 0.005694393 0.16580386
#>  [7,] 0.060472669 0.060472669 0.06047267
#>  [8,] 1.028811666 1.573881841 1.15631571
#>  [9,] 0.001145852 0.001145852 0.16403145
#> [10,] 0.148633887 0.148633887 0.14863389
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.90418165 0.184203330 0.54491797
#>  [2,] 0.19405383 0.194053826 0.01965369
#>  [3,] 0.08183524 0.081835243 0.08183524
#>  [4,] 0.08753074 0.745375248 0.44834413
#>  [5,] 1.75971570 2.106931479 2.10693148
#>  [6,] 0.02746793 0.302937614 0.30293761
#>  [7,] 0.34337663 0.343376630 0.28476859
#>  [8,] 0.12662757 1.116080136 0.10325290
#>  [9,] 0.39032099 0.404694525 0.39032099
#> [10,] 0.36601528 0.007427164 0.46637451
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.18391680 0.183916804 0.18391680
#>  [2,] 0.13738314 0.192957334 0.27913339
#>  [3,] 0.67151618 0.754569776 0.75456978
#>  [4,] 0.28325933 0.262042033 0.28325933
#>  [5,] 0.22285856 0.566099541 0.22285856
#>  [6,] 0.21158919 0.621353013 0.02481233
#>  [7,] 0.03121854 0.031218540 0.03121854
#>  [8,] 1.37994549 0.059926136 0.09258337
#>  [9,] 1.23492559 0.004712965 0.23389670
#> [10,] 0.24377091 0.243770906 0.24377091
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58164879 0.99207917 0.99207917
#>  [2,] 0.63928193 0.90673547 0.68562134
#>  [3,] 0.15516584 0.15516584 0.15516584
#>  [4,] 0.36758501 0.29832624 0.30224431
#>  [5,] 0.54972242 0.13195467 0.26353795
#>  [6,] 0.11090291 0.11090291 0.11090291
#>  [7,] 0.42373371 0.02253708 0.42373371
#>  [8,] 0.03136126 0.03136126 0.03136126
#>  [9,] 0.15297107 0.15297107 0.15297107
#> [10,] 0.23681492 0.03387064 0.44130350

## Inverse Gaussian
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14996336 0.45238476 0.14996336
#>  [2,] 3.23121864 0.03603193 0.16652141
#>  [3,] 0.10994811 0.31082460 0.31953307
#>  [4,] 0.31919461 0.44498620 0.44498620
#>  [5,] 0.20038485 0.20038485 0.18415632
#>  [6,] 0.23429198 0.58277763 0.67749312
#>  [7,] 0.17249691 0.17249691 0.17249691
#>  [8,] 0.65155131 0.65155131 0.65155131
#>  [9,] 0.11452062 0.11452062 0.15399316
#> [10,] 0.05678728 0.05678728 0.05678728
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.090890709 0.090890709 0.090890709
#>  [2,] 0.239472314 0.212320781 0.620136316
#>  [3,] 1.355028604 0.414149753 0.144894429
#>  [4,] 0.494964433 0.070128759 1.006316295
#>  [5,] 0.429662512 0.429662512 0.123757086
#>  [6,] 0.701317335 0.701317335 0.697282741
#>  [7,] 0.262350788 0.617875479 0.567822044
#>  [8,] 1.154573338 1.039831513 0.069012230
#>  [9,] 0.005186115 0.005186115 0.005186115
#> [10,] 0.674535042 0.150754952 0.236777095
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.192016934 0.2268595 0.1189842
#>  [2,] 0.532425361 0.9444067 0.3429348
#>  [3,] 0.003978442 0.1510824 0.1510824
#>  [4,] 0.014538354 0.3917283 0.3917283
#>  [5,] 0.094419797 0.0944198 0.1661945
#>  [6,] 0.231774285 0.5386194 0.2317743
#>  [7,] 0.147191973 0.1471920 0.1471920
#>  [8,] 0.172618379 0.1726184 0.3269301
#>  [9,] 0.133825254 0.1338253 0.1338253
#> [10,] 0.112247061 1.0090289 0.6900176
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58774642 1.08576060 0.83796484
#>  [2,] 0.16146565 0.17021280 0.57594964
#>  [3,] 0.61115948 0.14434835 0.01020577
#>  [4,] 0.53893596 0.59555427 0.41177556
#>  [5,] 0.16544211 0.16544211 0.39172493
#>  [6,] 0.38960491 0.40802798 0.07605635
#>  [7,] 0.02163386 0.08657338 0.17539024
#>  [8,] 0.13083021 0.18903598 0.25220726
#>  [9,] 0.33673336 0.03507220 0.59188408
#> [10,] 0.15592603 0.15592603 0.15592603

## Gamma
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31997889 0.87162836 0.31997889
#>  [2,] 0.72743113 0.15976870 0.70100793
#>  [3,] 0.25220308 0.25220308 0.08898803
#>  [4,] 0.72613250 0.16930955 0.72613250
#>  [5,] 0.44009534 0.31169924 0.97434027
#>  [6,] 0.16168848 0.03014465 0.51349764
#>  [7,] 0.03565084 0.03565084 0.03565084
#>  [8,] 0.02458440 0.02467084 0.11344432
#>  [9,] 0.92177898 0.85858386 0.21702892
#> [10,] 0.25389544 0.27874258 0.25389544
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.79795218 0.02581153 0.33823504
#>  [2,] 0.29551959 0.29551959 0.29551959
#>  [3,] 0.06756874 0.06756874 0.06756874
#>  [4,] 0.82728078 0.03617194 0.82728078
#>  [5,] 0.25158864 0.11097428 0.60180550
#>  [6,] 0.39982807 0.72882671 0.21276745
#>  [7,] 0.34680221 0.34680221 0.51381110
#>  [8,] 0.01183167 0.01183167 0.01183167
#>  [9,] 0.16597668 0.06261732 0.16597668
#> [10,] 0.00246560 0.12120050 0.00246560
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39207929 0.10635266 0.11882706
#>  [2,] 0.71211292 0.80344216 0.35491543
#>  [3,] 0.01091984 0.01091984 0.01091984
#>  [4,] 0.26990164 0.96991522 0.47836260
#>  [5,] 0.11800625 0.33662733 0.26313851
#>  [6,] 0.57975234 1.43861904 0.36025506
#>  [7,] 0.48942525 0.01139619 0.50326562
#>  [8,] 0.01248361 0.53191426 0.53191426
#>  [9,] 0.24732094 0.37754527 0.24732094
#> [10,] 0.40912753 0.26040642 0.63857834
rpextmo(
    10, 3, a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1439924 0.14399244 0.14399244
#>  [2,] 0.9988912 0.32447516 0.14916438
#>  [3,] 0.7915778 0.79157778 0.79157778
#>  [4,] 0.3111852 0.01201264 0.01201264
#>  [5,] 0.3891668 0.38916680 0.70677544
#>  [6,] 0.2538525 0.50098412 0.38113284
#>  [7,] 0.5909149 0.34788183 0.34788183
#>  [8,] 0.6263525 0.56046092 0.62635246
#>  [9,] 0.1677097 0.23257771 0.23257771
#> [10,] 0.2203931 0.49003699 1.23471514