Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13386272 3.08983521 1.17176611
#>  [2,] 5.68835187 1.28603107 1.72672659
#>  [3,] 0.14075590 0.14075590 0.14075590
#>  [4,] 0.16898757 0.16898757 0.16898757
#>  [5,] 2.05832504 2.11687791 2.11687791
#>  [6,] 0.01962169 0.01962169 0.01962169
#>  [7,] 0.73686335 0.64778367 0.54683624
#>  [8,] 0.90448682 0.65172332 0.15441468
#>  [9,] 1.47609749 0.22445646 0.83707600
#> [10,] 1.88881452 1.88881452 1.88881452
rpextmo(10, 3, a = 1)      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1639689 1.1639689 1.1639689
#>  [2,] 0.5978807 0.5978807 0.5978807
#>  [3,] 0.9695035 0.9695035 0.9695035
#>  [4,] 0.6321844 0.6321844 0.6321844
#>  [5,] 0.2630709 0.2630709 0.2630709
#>  [6,] 3.2939302 3.2939302 3.2939302
#>  [7,] 0.1354621 0.1354621 0.1354621
#>  [8,] 1.0098429 1.0098429 1.0098429
#>  [9,] 1.6934301 1.6934301 1.6934301
#> [10,] 0.6025583 0.6025583 0.6025583
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.07483822 0.5058906 1.1836169
#>  [2,] 0.67660787 0.3308783 0.5603389
#>  [3,] 1.16173032 0.1946235 0.7571237
#>  [4,] 1.23660619 0.1902697 0.7627902
#>  [5,] 0.19994572 0.5776944 0.2574009
#>  [6,] 0.15118827 2.0841770 0.1922339
#>  [7,] 1.16960482 0.2426706 0.9028067
#>  [8,] 2.47087001 1.5864683 1.4014065
#>  [9,] 0.04729294 0.8166686 0.3690182
#> [10,] 0.39984744 0.1162252 0.2474444

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3592114 0.3592114 0.3592114
#>  [2,] 0.5394398 0.5480373 0.5480373
#>  [3,] 1.0134051 1.0134051 0.4642262
#>  [4,] 1.7226825 1.7226825 1.7226825
#>  [5,] 6.4382244 4.8677579 0.4781806
#>  [6,] 0.4362816 1.0716851 1.0177794
#>  [7,] 2.2855014 2.1443562 0.9054308
#>  [8,] 0.7358717 7.3640385 1.2735884
#>  [9,] 0.8516855 0.5865154 1.7189149
#> [10,] 0.3590855 5.2252330 1.2152293
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.24247824 0.24247824 0.24247824
#>  [2,] 0.27225134 0.27225134 0.27225134
#>  [3,] 0.92062029 0.92062029 0.92062029
#>  [4,] 3.30041397 3.30041397 3.30041397
#>  [5,] 0.35240297 0.35240297 0.35240297
#>  [6,] 1.21231722 1.21231722 1.21231722
#>  [7,] 0.96950608 0.96950608 0.96950608
#>  [8,] 1.06520888 1.06520888 1.06520888
#>  [9,] 0.18238834 0.18238834 0.18238834
#> [10,] 0.06828455 0.06828455 0.06828455
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.52876058 0.64238545 3.6203874
#>  [2,] 1.77593460 3.28299408 0.5685051
#>  [3,] 1.10336210 0.48180630 0.1377355
#>  [4,] 0.02013937 0.36534588 1.9633553
#>  [5,] 0.48252230 0.19236006 4.8975011
#>  [6,] 0.28299987 0.53314643 0.2779766
#>  [7,] 1.79633045 0.91907011 0.8509018
#>  [8,] 0.37262878 0.06934391 0.2450019
#>  [9,] 1.01589418 0.26772249 2.8683700
#> [10,] 0.25501905 2.95536110 0.2667200

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.01938145 1.8838925 0.43417667
#>  [2,] 0.01520726 0.8478331 0.84783314
#>  [3,] 0.20119050 0.2011905 0.20119050
#>  [4,] 1.72700750 2.8796773 3.79573547
#>  [5,] 0.10053724 6.0097486 0.59279187
#>  [6,] 4.30447410 0.1846477 0.97758194
#>  [7,] 0.61611351 4.1575087 0.77888997
#>  [8,] 0.89085761 0.8908576 0.28673714
#>  [9,] 0.01728796 4.4456050 4.34508495
#> [10,] 1.73556639 1.8290355 0.04413985
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.469605861 0.469605861 0.469605861
#>  [2,] 0.030367432 0.030367432 0.030367432
#>  [3,] 1.328579870 1.328579870 1.328579870
#>  [4,] 1.445782054 1.445782054 1.445782054
#>  [5,] 2.608201625 2.608201625 2.608201625
#>  [6,] 2.481961449 2.481961449 2.481961449
#>  [7,] 1.740916900 1.740916900 1.740916900
#>  [8,] 0.627454445 0.627454445 0.627454445
#>  [9,] 1.768698601 1.768698601 1.768698601
#> [10,] 0.005598194 0.005598194 0.005598194
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.02010708 0.31974364 0.2921312
#>  [2,] 1.59034597 0.23402898 0.2667498
#>  [3,] 0.55947571 0.23311239 0.5219006
#>  [4,] 0.54371889 0.33150574 0.2202801
#>  [5,] 2.72236556 0.09531796 2.1679893
#>  [6,] 0.62602742 0.58374044 0.5192868
#>  [7,] 1.05178145 0.96382814 0.1400133
#>  [8,] 0.98809608 2.51213743 0.0996032
#>  [9,] 0.49755267 0.18150378 0.5021206
#> [10,] 0.73334479 0.98320082 0.7279121

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1848793 1.0345229 1.0345229
#>  [2,] 2.2752633 2.2752633 0.2145266
#>  [3,] 0.9532871 0.9532871 0.9532871
#>  [4,] 0.5096329 0.2127994 0.9392897
#>  [5,] 0.3081475 1.2605733 5.6508518
#>  [6,] 0.4678771 1.2359216 0.6347323
#>  [7,] 1.6275014 1.9140700 2.2292787
#>  [8,] 0.9131789 0.4519265 0.2665563
#>  [9,] 1.0509353 0.3135398 1.6709670
#> [10,] 1.0198123 2.6009768 0.9963189
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0101773 1.0101773 1.0101773
#>  [2,] 0.8742640 0.8742640 0.8742640
#>  [3,] 1.1984171 1.1984171 1.1984171
#>  [4,] 0.3533289 0.3533289 0.3533289
#>  [5,] 0.2252707 0.2252707 0.2252707
#>  [6,] 0.4734305 0.4734305 0.4734305
#>  [7,] 1.2529586 1.2529586 1.2529586
#>  [8,] 0.1644336 0.1644336 0.1644336
#>  [9,] 0.2263104 0.2263104 0.2263104
#> [10,] 0.3484298 0.3484298 0.3484298
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04935311 2.59288757 1.32281627
#>  [2,] 5.59828147 0.08656956 0.49668129
#>  [3,] 0.74994834 0.41864690 0.59959631
#>  [4,] 1.24131550 0.24410186 0.45792083
#>  [5,] 0.06289067 4.64813132 0.43309571
#>  [6,] 0.45805485 0.44352654 0.09343103
#>  [7,] 0.34915057 0.06700792 0.59322910
#>  [8,] 2.14524217 1.75544634 1.36548583
#>  [9,] 0.30063588 1.23717618 0.17956103
#> [10,] 0.04203566 0.43207367 0.14552890

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.88531253 0.8853125 0.8853125
#>  [2,] 0.43017564 0.6829858 0.6829858
#>  [3,] 3.60672885 2.3335152 1.2546727
#>  [4,] 1.57966538 1.5786717 0.6006030
#>  [5,] 2.64820957 5.5267768 1.0616791
#>  [6,] 1.70898872 0.5850540 0.6221489
#>  [7,] 0.06891728 2.5292619 2.5292619
#>  [8,] 8.60802923 2.9054161 0.5884285
#>  [9,] 0.10479637 1.1490560 1.1490560
#> [10,] 2.47094592 2.4709459 0.7455738
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4762420 0.4762420 0.4762420
#>  [2,] 1.1714387 1.1714387 1.1714387
#>  [3,] 1.1247321 1.1247321 1.1247321
#>  [4,] 1.6055013 1.6055013 1.6055013
#>  [5,] 0.2606451 0.2606451 0.2606451
#>  [6,] 2.1156643 2.1156643 2.1156643
#>  [7,] 3.8208535 3.8208535 3.8208535
#>  [8,] 2.8459525 2.8459525 2.8459525
#>  [9,] 0.1251910 0.1251910 0.1251910
#> [10,] 0.3574757 0.3574757 0.3574757
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.6191965 0.38101550 0.6368020
#>  [2,] 0.1987374 0.56442245 1.8825388
#>  [3,] 1.6415707 0.71028076 0.5827744
#>  [4,] 2.5028408 1.89388626 0.4235560
#>  [5,] 0.3379766 0.05695392 2.8833858
#>  [6,] 0.1980572 1.58716162 0.2554553
#>  [7,] 0.1366131 1.62859781 0.2013896
#>  [8,] 1.5409133 0.73524506 0.1772707
#>  [9,] 1.5979652 4.61173883 1.1176545
#> [10,] 2.4538771 1.58785903 1.1360505

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0978115 0.4304236 1.1181603
#>  [2,] 1.4763917 0.2563061 1.4763917
#>  [3,] 3.3870713 2.9659131 2.3000200
#>  [4,] 2.7278313 3.8463421 0.2645491
#>  [5,] 3.7142614 3.7142614 1.8191167
#>  [6,] 0.1332961 0.1332961 0.1332961
#>  [7,] 0.1801748 1.0373567 1.9334982
#>  [8,] 0.6528310 1.7526017 2.3642471
#>  [9,] 1.9934158 1.2309941 2.9882440
#> [10,] 1.0397491 2.6822018 2.2176194
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4013004 0.4013004 0.4013004
#>  [2,] 0.5609508 0.5609508 0.5609508
#>  [3,] 0.9283803 0.9283803 0.9283803
#>  [4,] 0.5577371 0.5577371 0.5577371
#>  [5,] 3.0933780 3.0933780 3.0933780
#>  [6,] 2.4807293 2.4807293 2.4807293
#>  [7,] 0.6200845 0.6200845 0.6200845
#>  [8,] 0.4335798 0.4335798 0.4335798
#>  [9,] 0.5625541 0.5625541 0.5625541
#> [10,] 1.0508103 1.0508103 1.0508103
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7168215 0.3874384 0.9253777
#>  [2,] 0.3354385 1.0792843 0.4138050
#>  [3,] 0.9963279 0.5250002 1.4514287
#>  [4,] 0.2289402 0.4754772 0.6676741
#>  [5,] 0.2294589 0.2331221 1.1862903
#>  [6,] 0.4417818 0.5737984 1.1005514
#>  [7,] 0.6864297 0.3955569 1.1706924
#>  [8,] 0.3148808 0.8974628 5.6897619
#>  [9,] 0.7124603 0.5778083 0.1908900
#> [10,] 0.3800622 2.8625642 0.5180607

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.53575721 2.94869269 0.6279680
#>  [2,] 2.27551291 1.02078170 2.8777848
#>  [3,] 0.31013573 0.07100173 1.5257904
#>  [4,] 0.42217747 0.42217747 0.4221775
#>  [5,] 0.05070972 0.79689916 1.9847228
#>  [6,] 0.48563549 0.35011144 0.5142082
#>  [7,] 0.23905157 0.81963951 4.7045800
#>  [8,] 0.18315198 0.85968626 0.5734035
#>  [9,] 1.71829681 1.71829681 0.6367772
#> [10,] 3.09897833 0.42294429 1.3915646
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.71557711 1.71557711 1.71557711
#>  [2,] 0.68776089 0.68776089 0.68776089
#>  [3,] 1.39637575 1.39637575 1.39637575
#>  [4,] 0.97099362 0.97099362 0.97099362
#>  [5,] 0.77336097 0.77336097 0.77336097
#>  [6,] 0.91737683 0.91737683 0.91737683
#>  [7,] 0.07553835 0.07553835 0.07553835
#>  [8,] 0.28495841 0.28495841 0.28495841
#>  [9,] 2.26634221 2.26634221 2.26634221
#> [10,] 1.64719102 1.64719102 1.64719102
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.34408273 3.4559484 0.7797914
#>  [2,] 0.52233148 1.1458364 0.2305758
#>  [3,] 0.59096146 0.2418350 1.7643887
#>  [4,] 0.64727692 1.2149687 3.4364894
#>  [5,] 0.06981989 0.9639769 0.9387771
#>  [6,] 0.26473501 0.6413367 0.5992910
#>  [7,] 0.03256951 1.4815822 2.7428659
#>  [8,] 0.85622223 0.7130015 0.8361474
#>  [9,] 0.47291242 1.3616654 0.4257135
#> [10,] 0.80362837 0.9304927 0.0775922

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]      [,2]        [,3]
#>  [1,] 3.1336561 1.3572500 1.658667702
#>  [2,] 2.9644325 1.8327486 3.202318154
#>  [3,] 1.5844086 1.1108543 1.985506006
#>  [4,] 1.2699861 0.5385358 1.896109216
#>  [5,] 0.7768983 0.3616746 0.776898348
#>  [6,] 1.2211638 2.9941175 0.004010399
#>  [7,] 0.9169369 0.9169369 0.916936891
#>  [8,] 0.6459530 1.8134937 2.321498205
#>  [9,] 0.9642067 1.9621836 2.315103265
#> [10,] 0.4832782 0.4832782 0.073508800
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4382288 0.4382288 0.4382288
#>  [2,] 0.3942656 0.3942656 0.3942656
#>  [3,] 0.3875019 0.3875019 0.3875019
#>  [4,] 0.9828898 0.9828898 0.9828898
#>  [5,] 0.3109049 0.3109049 0.3109049
#>  [6,] 1.9576167 1.9576167 1.9576167
#>  [7,] 0.1668756 0.1668756 0.1668756
#>  [8,] 1.0632076 1.0632076 1.0632076
#>  [9,] 0.2672686 0.2672686 0.2672686
#> [10,] 0.6730612 0.6730612 0.6730612
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]        [,3]
#>  [1,] 1.73174898 0.1666504 3.489167089
#>  [2,] 1.29669593 0.1868110 2.266886683
#>  [3,] 0.33857624 1.2305994 0.484755384
#>  [4,] 3.10962300 1.1630970 1.133435888
#>  [5,] 0.08388507 0.6878982 0.688544904
#>  [6,] 0.51378590 0.8008050 1.974121486
#>  [7,] 1.87047557 1.0865197 0.006539822
#>  [8,] 0.63596436 1.6353352 0.252475910
#>  [9,] 0.17666198 1.5300686 0.792760804
#> [10,] 0.19551284 0.4670655 0.115531118

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.5968633 1.5203359 5.3669525
#>  [2,] 0.2664371 0.2786002 0.3580206
#>  [3,] 0.6506515 0.3522236 8.5029940
#>  [4,] 2.3155717 2.3155717 2.3155717
#>  [5,] 1.9339990 0.9566814 2.2489970
#>  [6,] 0.3496807 3.6342162 3.6205397
#>  [7,] 2.7540509 0.1735506 0.9082075
#>  [8,] 0.8243526 0.1061804 0.4378632
#>  [9,] 0.6044968 2.1824904 1.3607737
#> [10,] 1.3666475 3.5879691 0.3915715
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 3.343482941 3.343482941 3.343482941
#>  [2,] 0.629391786 0.629391786 0.629391786
#>  [3,] 0.001121206 0.001121206 0.001121206
#>  [4,] 0.196411361 0.196411361 0.196411361
#>  [5,] 2.531842338 2.531842338 2.531842338
#>  [6,] 0.432347094 0.432347094 0.432347094
#>  [7,] 0.518596622 0.518596622 0.518596622
#>  [8,] 0.260741484 0.260741484 0.260741484
#>  [9,] 1.209231228 1.209231228 1.209231228
#> [10,] 2.418882819 2.418882819 2.418882819
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 2.50985873 1.04893510 2.7451922
#>  [2,] 2.21673762 0.22056038 2.8852571
#>  [3,] 0.81875916 0.59852156 0.3760267
#>  [4,] 0.05588930 0.75061410 1.2125315
#>  [5,] 0.70148913 0.61599401 0.6325394
#>  [6,] 0.40101682 3.03775204 0.1536161
#>  [7,] 1.36533336 3.83529595 0.4979077
#>  [8,] 0.02927765 2.63866404 0.2967067
#>  [9,] 0.28613628 2.65153460 4.7774256
#> [10,] 0.30397778 0.02738456 0.5806131

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2967171 0.66681045 1.4706563
#>  [2,] 4.3184191 0.26254145 0.8609257
#>  [3,] 0.6836484 0.86438377 3.0437524
#>  [4,] 0.4364913 1.90310605 0.7496040
#>  [5,] 0.1570159 0.34823264 1.3679762
#>  [6,] 4.7296718 0.07811675 2.6601295
#>  [7,] 1.7963861 1.62589767 0.4730390
#>  [8,] 1.7875644 0.95419720 1.9979189
#>  [9,] 0.7954428 1.31561903 1.3156190
#> [10,] 0.2783799 0.43968757 0.3867574
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09280649 0.09280649 0.09280649
#>  [2,] 1.73083648 1.73083648 1.73083648
#>  [3,] 1.79878345 1.79878345 1.79878345
#>  [4,] 1.19137716 1.19137716 1.19137716
#>  [5,] 2.17374042 2.17374042 2.17374042
#>  [6,] 1.23524754 1.23524754 1.23524754
#>  [7,] 1.04058042 1.04058042 1.04058042
#>  [8,] 1.39555095 1.39555095 1.39555095
#>  [9,] 0.32540168 0.32540168 0.32540168
#> [10,] 1.38335234 1.38335234 1.38335234
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1765412 0.69535865 0.76620948
#>  [2,] 0.2980808 0.51264796 1.35164360
#>  [3,] 0.7635065 0.61138878 0.03854057
#>  [4,] 0.7864486 2.47539564 0.04673785
#>  [5,] 0.2776933 0.15274009 0.56932269
#>  [6,] 0.8576676 1.01412726 0.19428212
#>  [7,] 0.3669628 0.05731569 1.28127276
#>  [8,] 1.0770432 0.02610849 0.28707511
#>  [9,] 0.1574961 0.29479867 0.09965929
#> [10,] 0.2030354 0.91051879 0.11151891

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.41324907 0.1757943 0.4132491
#>  [2,] 0.33368048 0.2197719 0.3336805
#>  [3,] 0.32217504 0.3221750 0.8515836
#>  [4,] 3.60100772 2.4568211 1.4413699
#>  [5,] 0.75107125 0.1899367 1.0682746
#>  [6,] 2.03528682 0.5841640 0.3272150
#>  [7,] 0.99821353 0.1355292 0.1355292
#>  [8,] 0.20030333 0.4322684 0.4322684
#>  [9,] 0.49797873 1.3618795 2.7890389
#> [10,] 0.09045527 1.6215167 3.0414460
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05667741 0.26180594 0.05667741
#>  [2,] 1.66184847 0.08075242 1.51238783
#>  [3,] 0.27638096 0.23638842 1.36154029
#>  [4,] 3.03110891 1.83555761 0.75955998
#>  [5,] 0.83834038 0.83834038 0.23260648
#>  [6,] 0.94085841 1.82228211 0.06010974
#>  [7,] 1.70514670 2.65836293 1.02146653
#>  [8,] 0.58140279 0.19572186 1.66742749
#>  [9,] 1.08404719 0.10521094 0.15881709
#> [10,] 1.08735069 0.15396525 0.15396525
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.10339251 2.44120911 0.0246289
#>  [2,] 0.27751325 0.39487220 1.4735588
#>  [3,] 3.19462427 0.40149048 3.1946243
#>  [4,] 0.29098037 0.36883975 0.3688397
#>  [5,] 0.06742433 0.38135231 0.2723335
#>  [6,] 0.09379506 0.44777846 0.3162271
#>  [7,] 0.02045919 0.02045919 0.1240824
#>  [8,] 0.33768802 1.68713220 0.3376880
#>  [9,] 0.44060075 0.12883178 1.4188101
#> [10,] 0.01366361 1.44581947 1.5737245
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32673858 0.60606184 0.32673858
#>  [2,] 0.08686977 1.49330301 0.51079910
#>  [3,] 0.24031870 0.13513415 0.32518410
#>  [4,] 0.20985915 0.24796611 0.15058965
#>  [5,] 0.25771092 0.25771092 1.44201390
#>  [6,] 0.27782150 0.27782150 0.27782150
#>  [7,] 0.39360615 0.14992335 0.51454104
#>  [8,] 0.09749573 0.09749573 0.09749573
#>  [9,] 1.06807976 0.24195431 0.69413563
#> [10,] 0.52309986 0.45428409 0.05590687
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.10833869 0.4524997 0.33117610
#>  [2,] 0.93172481 0.6698274 0.26534002
#>  [3,] 1.45859043 3.8842261 1.18876174
#>  [4,] 1.18975369 0.5133351 0.58464341
#>  [5,] 0.31317723 0.2704858 1.01562880
#>  [6,] 0.85577383 0.5740831 0.85215374
#>  [7,] 1.58191800 0.9020781 1.37415035
#>  [8,] 0.04551995 1.0011745 0.04551995
#>  [9,] 0.42174978 0.1082685 1.22761694
#> [10,] 0.25444800 0.9232931 1.37440167

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]      [,2]      [,3]
#>  [1,] 4.93484584 0.7946074 3.1251106
#>  [2,] 2.65817428 3.6957097 4.7063137
#>  [3,] 5.10032317 0.5162260 0.3184454
#>  [4,] 0.54341868 0.5434187 0.5434187
#>  [5,] 0.47092093 0.4709209 0.4709209
#>  [6,] 0.43242021 0.5082516 0.4451888
#>  [7,] 0.05918275 2.5223001 3.5816317
#>  [8,] 0.02345950 1.6240669 3.4407952
#>  [9,] 0.02580727 0.8563233 0.3177425
#> [10,] 0.20547775 0.5166582 0.3473336
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6106944 1.3631821 0.4930779
#>  [2,] 3.3548036 1.0647071 3.3548036
#>  [3,] 0.1148144 0.1148144 0.1148144
#>  [4,] 0.6805185 0.3893364 0.7041310
#>  [5,] 4.4280763 0.2718665 0.4688096
#>  [6,] 1.3811159 0.2868396 1.3811159
#>  [7,] 1.4519497 1.4614199 2.5161094
#>  [8,] 1.7388349 1.7388349 1.0529924
#>  [9,] 2.0919037 1.3547300 4.9593741
#> [10,] 0.4507498 0.8827409 0.8827409
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.36317641 2.5086948 2.5086948
#>  [2,] 0.24906455 0.2490646 0.2490646
#>  [3,] 1.05415480 2.3123494 0.9059927
#>  [4,] 0.76124679 0.1872594 0.6314370
#>  [5,] 1.39917954 2.3087463 1.7027285
#>  [6,] 0.80895813 0.8089581 0.6342985
#>  [7,] 0.26558006 2.9670410 0.9580781
#>  [8,] 2.01078781 2.4240189 0.4101172
#>  [9,] 0.02769093 0.4185196 1.1142451
#> [10,] 0.04949705 4.3809603 4.3294389
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4438919 0.1926739 1.1786340
#>  [2,] 2.8511761 2.8511761 2.8511761
#>  [3,] 0.3413407 1.0606588 6.7259419
#>  [4,] 1.9441979 0.2399411 0.8193510
#>  [5,] 1.0785293 1.0785293 1.0785293
#>  [6,] 2.4181821 2.4181821 0.1188049
#>  [7,] 0.6325760 0.6325760 0.6325760
#>  [8,] 2.3630609 0.3044657 0.1341533
#>  [9,] 0.3640608 1.4762892 1.1769776
#> [10,] 2.6436333 2.6436333 0.3354036
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 3.6446054 3.49260062 1.0888271
#>  [2,] 1.2519845 0.89742345 3.5560790
#>  [3,] 5.4585334 0.44087097 5.3934460
#>  [4,] 0.4563417 0.45634168 0.3715399
#>  [5,] 2.4655032 1.77009675 0.8085564
#>  [6,] 0.1138768 0.51614449 0.3061925
#>  [7,] 0.1854459 0.25202164 0.2520216
#>  [8,] 0.3160689 0.05386761 0.3160689
#>  [9,] 2.0867385 0.94705571 1.4631911
#> [10,] 3.7551876 0.52724370 1.4258837

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18196684 0.18196684 0.18196684
#>  [2,] 0.28558135 0.28558135 0.60261679
#>  [3,] 0.02508249 0.02508249 0.02508249
#>  [4,] 0.50665678 0.50665678 0.34164615
#>  [5,] 0.42467039 1.97548778 0.93242362
#>  [6,] 0.75361066 0.75361066 0.51728122
#>  [7,] 0.14239696 0.19543203 0.19543203
#>  [8,] 0.01911629 0.01911629 0.01911629
#>  [9,] 0.64042674 0.64042674 1.35581483
#> [10,] 0.11430376 0.11430376 0.11430376
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.52073541 0.06128998 0.52073541
#>  [2,] 0.06175609 0.06175609 0.06175609
#>  [3,] 0.32953400 0.32953400 0.32953400
#>  [4,] 1.90779541 2.01160698 1.90779541
#>  [5,] 0.62490931 0.62490931 0.62490931
#>  [6,] 0.05268152 0.05268152 0.05268152
#>  [7,] 1.07049192 0.22345873 0.22345873
#>  [8,] 0.84786302 2.58150176 0.97643372
#>  [9,] 1.46609424 1.30176180 2.82370654
#> [10,] 0.69021936 0.69021936 0.69021936
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3883376 0.02567147 0.02567147
#>  [2,] 0.2160528 0.96336026 0.21994935
#>  [3,] 0.1371415 0.13714153 0.13714153
#>  [4,] 1.1831326 0.16523570 0.26817642
#>  [5,] 0.6470494 0.56655606 0.64704936
#>  [6,] 0.5315925 0.53159254 0.54221523
#>  [7,] 0.2258250 0.22582504 0.22582504
#>  [8,] 0.4048377 0.79028995 0.79028995
#>  [9,] 0.7450709 0.69299689 0.85930880
#> [10,] 1.0394782 0.94764244 1.03947823
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.15386047 0.5581711 0.10945649
#>  [2,] 0.16306924 2.2907070 0.16306924
#>  [3,] 0.22220025 0.2728634 0.27286336
#>  [4,] 0.06725521 0.3410713 0.05467175
#>  [5,] 0.36307282 0.3341811 1.18630137
#>  [6,] 0.13474726 0.1347473 0.13474726
#>  [7,] 0.39093880 0.1350679 0.39093880
#>  [8,] 0.13830324 0.1383032 0.13830324
#>  [9,] 0.02491819 0.2797894 0.27978942
#> [10,] 0.39819118 0.3981912 0.39819118
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2360412 1.2360412 0.23526216
#>  [2,] 0.8291772 0.9562149 0.11230963
#>  [3,] 0.3720714 0.1164372 0.79412236
#>  [4,] 1.3397590 0.5955688 0.36770364
#>  [5,] 0.2021701 0.2021701 0.13137382
#>  [6,] 0.1114385 0.1114385 0.01800015
#>  [7,] 0.2275708 0.2275708 0.05666046
#>  [8,] 0.5099873 1.4170544 1.01404430
#>  [9,] 0.5812776 1.5517779 1.54256962
#> [10,] 0.1997853 0.9721660 0.78957804

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11558675 0.11558675 0.11558675
#>  [2,] 1.09901973 1.45639841 0.62572517
#>  [3,] 0.22873647 0.22873647 0.58554207
#>  [4,] 0.70309589 0.74075199 1.68632472
#>  [5,] 0.21120757 0.21120757 0.21120757
#>  [6,] 0.37256323 0.41424081 0.25999867
#>  [7,] 0.05473711 0.64297318 0.65143935
#>  [8,] 0.97251341 0.23502891 0.23502891
#>  [9,] 1.10719128 0.04137044 0.16006191
#> [10,] 0.04344473 0.04344473 0.04344473
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.21919472 0.180678451 0.22693894
#>  [2,] 0.26409593 0.285157679 0.01122598
#>  [3,] 2.33976480 0.790873632 0.79087363
#>  [4,] 0.01755289 0.017552891 0.01755289
#>  [5,] 0.94117330 0.001351038 0.22403449
#>  [6,] 0.52459984 0.524599836 0.12058285
#>  [7,] 0.30346127 0.303461266 0.30346127
#>  [8,] 0.17456472 0.174564722 0.17456472
#>  [9,] 0.81513483 0.530142255 0.30821079
#> [10,] 0.06686707 0.986882394 0.38586085
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02882172 0.02882172 0.02882172
#>  [2,] 1.14605064 0.62135877 1.00898724
#>  [3,] 1.20137501 1.74771093 1.74771093
#>  [4,] 0.24963736 0.24963736 0.05370438
#>  [5,] 0.13116313 0.81654251 0.90696633
#>  [6,] 0.56592968 0.07270611 0.56592968
#>  [7,] 0.23180263 0.23180263 0.05155147
#>  [8,] 0.40466681 0.14020752 0.40466681
#>  [9,] 1.24074306 1.16006883 0.00209017
#> [10,] 0.03521192 0.70801482 0.06423896
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05098854 0.05098854 0.05098854
#>  [2,] 0.63720373 0.49497397 0.33606088
#>  [3,] 1.11764566 0.04644113 0.08036366
#>  [4,] 0.05018293 0.17231808 0.17231808
#>  [5,] 0.33911467 0.01962916 0.18517207
#>  [6,] 0.20620837 0.16059092 0.16059092
#>  [7,] 0.12527597 0.12527597 0.12527597
#>  [8,] 1.33880827 0.92271403 0.29942429
#>  [9,] 0.26713310 0.10243737 0.67290029
#> [10,] 0.20860927 0.20860927 0.20860927

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05604068 0.51716137 0.03139305
#>  [2,] 0.08837724 0.09235649 0.13796962
#>  [3,] 0.15529676 0.15529676 0.15529676
#>  [4,] 0.39637295 0.73387718 0.02663044
#>  [5,] 0.60580429 0.37112767 0.58058513
#>  [6,] 0.17231703 0.17231703 0.38651150
#>  [7,] 1.24427774 0.15721843 0.09763975
#>  [8,] 0.11721011 0.02449228 0.09182127
#>  [9,] 0.10111906 0.17648925 0.11992494
#> [10,] 0.10047716 0.82804624 0.01068554
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.33947063 0.04282406 0.042824061
#>  [2,] 0.51157166 0.66333738 0.663337381
#>  [3,] 0.09887673 0.40491236 0.404912358
#>  [4,] 0.27761801 0.73037584 0.718695966
#>  [5,] 0.08646064 0.05504856 0.001931599
#>  [6,] 0.92127852 0.92127852 0.921278521
#>  [7,] 0.28461588 0.10397885 0.284615879
#>  [8,] 0.58857559 0.58857559 0.492427012
#>  [9,] 0.06011097 0.06011097 0.060110971
#> [10,] 0.40464903 0.40464903 0.986686311
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38681611 0.38681611 0.38681611
#>  [2,] 0.73969807 0.11081812 0.26130623
#>  [3,] 0.17962452 0.24223857 0.04007394
#>  [4,] 0.05111678 0.05111678 0.13768198
#>  [5,] 0.43600261 0.01329794 0.75290954
#>  [6,] 0.20140031 0.80430463 0.28184622
#>  [7,] 0.11958306 1.12285384 0.01254620
#>  [8,] 0.26507793 0.21506677 0.26507793
#>  [9,] 0.01391988 0.19539311 0.19539311
#> [10,] 0.57154699 0.57154699 0.33464327
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.7387515 0.93542907 0.43702323
#>  [2,] 0.2097215 0.19239558 0.20972155
#>  [3,] 0.8423821 0.84238213 0.43649889
#>  [4,] 0.2700695 0.07180369 0.06079595
#>  [5,] 0.1111145 0.34476567 0.34476567
#>  [6,] 0.5030025 0.03158771 0.03158771
#>  [7,] 0.4542341 0.24404932 1.79431905
#>  [8,] 0.1495564 0.14955640 0.14955640
#>  [9,] 0.1938752 0.19387520 0.19387520
#> [10,] 0.7348501 0.02279844 0.29248421

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.80532435 0.53426013 0.208669166
#>  [2,] 0.07138032 0.07138032 0.202595138
#>  [3,] 0.55638020 1.31138749 0.556380197
#>  [4,] 0.42823658 0.70873392 0.085163892
#>  [5,] 0.19133161 0.29940703 0.191331606
#>  [6,] 0.63232645 0.31477866 0.412990075
#>  [7,] 0.05733644 0.05733644 0.057336443
#>  [8,] 0.27626019 1.36640054 0.087621474
#>  [9,] 0.02071497 0.02071497 0.005550996
#> [10,] 0.11984308 1.08150614 1.502889805
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>            [,1]      [,2]        [,3]
#>  [1,] 0.3454396 0.3454396 0.008352804
#>  [2,] 0.2999466 0.8354735 0.714468727
#>  [3,] 0.3099945 0.2226962 0.222696228
#>  [4,] 0.2426244 0.4224277 0.539955039
#>  [5,] 0.3391525 0.4742046 0.339152490
#>  [6,] 0.0730986 0.0730986 0.073098601
#>  [7,] 0.0678880 0.0678880 0.067888004
#>  [8,] 0.1702391 0.8744774 0.053793749
#>  [9,] 0.5284955 1.2524668 0.885221072
#> [10,] 2.6547901 0.2729457 1.376801680
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>               [,1]         [,2]         [,3]
#>  [1,] 0.5211820892 0.9261594122 0.5211820892
#>  [2,] 0.2048095653 0.2048095653 0.2048095653
#>  [3,] 0.4687333018 0.4391871976 0.3067039050
#>  [4,] 0.0546893524 0.5628357635 1.2417103033
#>  [5,] 0.0495298717 0.0495298717 1.0648094595
#>  [6,] 0.1376164145 0.0468525398 0.1376164145
#>  [7,] 0.0984537666 0.5346735616 0.5346735616
#>  [8,] 0.0295676825 0.7285033390 0.1142809846
#>  [9,] 0.0008733226 0.0008733226 0.0008733226
#> [10,] 0.1940338945 0.3030694106 0.1940338945
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.62821703 0.19276747 0.36377269
#>  [2,] 1.81080951 0.51519671 0.43531392
#>  [3,] 0.02076083 0.02076083 0.02076083
#>  [4,] 0.15267614 0.07681584 0.18346346
#>  [5,] 0.05056111 0.08723351 0.44761691
#>  [6,] 0.45860206 0.13141430 0.45860206
#>  [7,] 0.03074543 0.03074543 0.03074543
#>  [8,] 0.06682502 0.06682502 0.06682502
#>  [9,] 0.05035569 0.61084225 0.13132195
#> [10,] 0.26992180 0.26992180 0.85469431