Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.48892860 2.55529183 1.71318375
#>  [2,] 3.36591473 1.20358011 1.92350952
#>  [3,] 1.74009543 1.06638490 2.92816681
#>  [4,] 4.74023778 5.29562535 1.24991473
#>  [5,] 1.00621024 1.28545537 0.01290057
#>  [6,] 0.51251647 2.63068520 2.63068520
#>  [7,] 5.79814228 0.02592406 4.01019227
#>  [8,] 2.72608041 0.42443567 1.50161301
#>  [9,] 0.06524833 1.07552164 0.53854701
#> [10,] 0.87342818 1.59295021 2.81535724
rpextmo(10, 3, a = 1)      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6426375 0.6426375 0.6426375
#>  [2,] 0.6704587 0.6704587 0.6704587
#>  [3,] 2.2342799 2.2342799 2.2342799
#>  [4,] 2.3215371 2.3215371 2.3215371
#>  [5,] 0.3036467 0.3036467 0.3036467
#>  [6,] 1.6295157 1.6295157 1.6295157
#>  [7,] 0.1525409 0.1525409 0.1525409
#>  [8,] 1.2222082 1.2222082 1.2222082
#>  [9,] 0.5705257 0.5705257 0.5705257
#> [10,] 0.4499137 0.4499137 0.4499137
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.22295576 0.05532225 0.35851796
#>  [2,] 0.08082859 0.87727201 0.84515687
#>  [3,] 0.55880323 1.01043837 1.44854814
#>  [4,] 0.21177696 0.78342997 0.92084358
#>  [5,] 0.15758199 0.46138513 1.27254491
#>  [6,] 0.50529494 4.48832002 0.58782823
#>  [7,] 1.40251165 1.73913150 0.83440650
#>  [8,] 1.76663449 0.23976755 0.16417841
#>  [9,] 0.93938299 0.48464396 0.03469087
#> [10,] 0.03983926 0.72762170 2.07208818

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.4672469 0.06599181 0.46724689
#>  [2,] 1.3634080 1.03421427 1.61005908
#>  [3,] 3.5561110 0.35376503 1.11925656
#>  [4,] 1.8923324 0.64518560 0.01068941
#>  [5,] 0.5241091 1.00112087 1.73164471
#>  [6,] 1.2337782 0.40999069 1.23377816
#>  [7,] 2.9468661 1.67482617 3.52689669
#>  [8,] 0.6915221 0.69152207 0.69152207
#>  [9,] 2.2312944 2.21358438 2.23129444
#> [10,] 0.5239326 0.52393259 0.52393259
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05837463 0.05837463 0.05837463
#>  [2,] 0.12621989 0.12621989 0.12621989
#>  [3,] 1.37406507 1.37406507 1.37406507
#>  [4,] 0.82009893 0.82009893 0.82009893
#>  [5,] 0.84620809 0.84620809 0.84620809
#>  [6,] 0.65901368 0.65901368 0.65901368
#>  [7,] 0.80997271 0.80997271 0.80997271
#>  [8,] 0.14276774 0.14276774 0.14276774
#>  [9,] 0.85627158 0.85627158 0.85627158
#> [10,] 0.07462592 0.07462592 0.07462592
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>            [,1]       [,2]        [,3]
#>  [1,] 1.0832551 0.25471160 0.486913927
#>  [2,] 0.3947531 0.33960889 1.214823006
#>  [3,] 0.4077766 0.11964526 1.128492443
#>  [4,] 0.0397764 2.21859016 0.625719896
#>  [5,] 0.3972933 1.86146566 0.082446673
#>  [6,] 0.8515295 0.68810777 1.804197360
#>  [7,] 3.3976632 0.41892238 0.007351023
#>  [8,] 0.8386387 2.47551521 1.508930586
#>  [9,] 1.4400206 1.16869243 0.505551068
#> [10,] 0.8860569 0.08308233 0.393036595

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 6.5847702 1.5257974 3.01748930
#>  [2,] 1.7706197 4.0790591 0.08144606
#>  [3,] 0.7536616 0.2545875 2.03801033
#>  [4,] 2.5005409 2.8571630 0.71179802
#>  [5,] 1.7031191 0.6079438 2.79246025
#>  [6,] 0.0119720 0.3852016 0.86356515
#>  [7,] 1.3840557 0.4967509 3.81937382
#>  [8,] 0.9523990 0.8249336 2.55534582
#>  [9,] 0.1380837 0.1380837 0.13808373
#> [10,] 1.3599333 0.9135155 0.84424228
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0611100 1.0611100 1.0611100
#>  [2,] 0.5010452 0.5010452 0.5010452
#>  [3,] 1.1812708 1.1812708 1.1812708
#>  [4,] 3.0691843 3.0691843 3.0691843
#>  [5,] 0.3794715 0.3794715 0.3794715
#>  [6,] 0.3205566 0.3205566 0.3205566
#>  [7,] 1.3789296 1.3789296 1.3789296
#>  [8,] 0.1303155 0.1303155 0.1303155
#>  [9,] 2.4690307 2.4690307 2.4690307
#> [10,] 6.2765562 6.2765562 6.2765562
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2280248 0.1405601 0.4393281
#>  [2,] 0.5134734 0.4152405 0.5063667
#>  [3,] 0.8668950 0.2753147 1.5486687
#>  [4,] 0.1710641 1.2706103 1.2298378
#>  [5,] 1.1514255 0.1526048 0.5498542
#>  [6,] 0.6919852 1.1157787 0.1945473
#>  [7,] 0.2691093 2.8090045 1.0031967
#>  [8,] 4.0362025 0.4127917 0.7287403
#>  [9,] 0.9710907 1.2427116 4.1491541
#> [10,] 4.1631620 0.5824678 0.4956268

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4075651 1.4075651 0.4406120
#>  [2,] 6.5757301 0.4557974 0.5278214
#>  [3,] 1.4128551 2.0875343 2.9996540
#>  [4,] 1.1980931 1.4824458 1.1699703
#>  [5,] 0.7180224 2.3030769 0.2380569
#>  [6,] 3.5326084 6.4297902 2.3297668
#>  [7,] 1.3744110 0.9175511 0.1737112
#>  [8,] 2.3193986 2.3193986 2.3193986
#>  [9,] 0.7013947 3.2345573 0.0540973
#> [10,] 0.2756449 0.2756449 0.2756449
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.66766291 0.66766291 0.66766291
#>  [2,] 0.93911935 0.93911935 0.93911935
#>  [3,] 1.37286363 1.37286363 1.37286363
#>  [4,] 0.07226139 0.07226139 0.07226139
#>  [5,] 0.96834675 0.96834675 0.96834675
#>  [6,] 0.35080761 0.35080761 0.35080761
#>  [7,] 0.64780562 0.64780562 0.64780562
#>  [8,] 0.22053503 0.22053503 0.22053503
#>  [9,] 0.04712611 0.04712611 0.04712611
#> [10,] 3.68319224 3.68319224 3.68319224
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 3.2424300 1.32801409 0.80538581
#>  [2,] 0.1254647 1.23125433 0.51851828
#>  [3,] 1.3950211 1.53058341 0.82890399
#>  [4,] 0.9874894 1.27757062 2.29369829
#>  [5,] 0.3433640 0.59656348 0.33461878
#>  [6,] 0.5851104 0.07149416 0.07393324
#>  [7,] 0.9048110 0.69441377 0.58022157
#>  [8,] 1.4323400 0.11275320 0.24795495
#>  [9,] 1.8982178 0.84418555 1.64452348
#> [10,] 0.2668138 1.35173148 0.84565429

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6439425 0.51635514 0.83088380
#>  [2,] 0.1205895 3.30829885 3.33701873
#>  [3,] 1.9893894 2.07055275 0.52129544
#>  [4,] 0.7209280 0.72092798 0.72092798
#>  [5,] 0.3626265 1.72533043 0.98066123
#>  [6,] 0.1378686 3.25616278 0.31172134
#>  [7,] 4.3281799 4.51641080 3.20600456
#>  [8,] 1.0498020 0.35615823 0.49935502
#>  [9,] 0.2314575 0.09359105 0.04780149
#> [10,] 0.2340824 4.20899434 1.36196642
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.82155953 0.82155953 0.82155953
#>  [2,] 2.11771318 2.11771318 2.11771318
#>  [3,] 0.04388746 0.04388746 0.04388746
#>  [4,] 0.19972378 0.19972378 0.19972378
#>  [5,] 0.71425702 0.71425702 0.71425702
#>  [6,] 3.16614088 3.16614088 3.16614088
#>  [7,] 0.68483833 0.68483833 0.68483833
#>  [8,] 2.19414491 2.19414491 2.19414491
#>  [9,] 0.57668238 0.57668238 0.57668238
#> [10,] 1.92209477 1.92209477 1.92209477
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.42905143 0.8799016 0.3721739
#>  [2,] 0.47261853 1.7329007 0.8446430
#>  [3,] 0.51724780 1.0228622 0.4600255
#>  [4,] 0.79975756 0.1098306 1.1053093
#>  [5,] 0.47765025 0.3028306 1.6395365
#>  [6,] 0.62944592 1.3863089 0.3417693
#>  [7,] 0.05413297 0.4535311 0.5850659
#>  [8,] 0.03130999 0.4260998 0.1458119
#>  [9,] 2.14027846 0.1153478 1.8582854
#> [10,] 3.86012103 0.1336243 1.2871054

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]       [,2]        [,3]
#>  [1,] 1.65126818 1.54670565 0.053521104
#>  [2,] 3.36485186 3.36485186 2.782448996
#>  [3,] 1.21781655 1.21781655 1.217816551
#>  [4,] 0.75074882 0.75074882 0.544561588
#>  [5,] 0.64208610 6.25402106 4.935895625
#>  [6,] 1.56503657 1.01187333 0.269650440
#>  [7,] 1.17363639 1.17363639 0.933691729
#>  [8,] 0.16047329 7.18348871 3.183701001
#>  [9,] 0.07831373 0.07831373 0.078313731
#> [10,] 0.49201072 1.50373532 0.001883435
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42854176 0.42854176 0.42854176
#>  [2,] 0.24312564 0.24312564 0.24312564
#>  [3,] 0.38171769 0.38171769 0.38171769
#>  [4,] 1.26389226 1.26389226 1.26389226
#>  [5,] 0.57879840 0.57879840 0.57879840
#>  [6,] 0.84478853 0.84478853 0.84478853
#>  [7,] 1.84526810 1.84526810 1.84526810
#>  [8,] 0.43342449 0.43342449 0.43342449
#>  [9,] 0.33170978 0.33170978 0.33170978
#> [10,] 0.01936096 0.01936096 0.01936096
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 3.14810216 0.9023053 0.2798798
#>  [2,] 0.73424843 1.6657612 2.0755797
#>  [3,] 0.50238307 0.9542007 0.4158884
#>  [4,] 0.07816511 0.4012132 0.6869097
#>  [5,] 0.60412903 1.6520859 0.4601373
#>  [6,] 0.53465631 4.0199592 0.0773431
#>  [7,] 2.09680640 0.6183693 0.1657093
#>  [8,] 0.07478958 1.2999213 0.2622561
#>  [9,] 0.45501502 0.6122993 1.8687633
#> [10,] 2.26072271 1.6350910 0.1391640

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>              [,1]       [,2]      [,3]
#>  [1,] 0.006469576 2.91358650 1.9561525
#>  [2,] 0.074777073 0.57689644 0.4864817
#>  [3,] 1.190333337 0.29161408 1.4267987
#>  [4,] 0.276180201 0.05548894 0.2761802
#>  [5,] 0.500708912 2.12282237 0.4829512
#>  [6,] 0.238489747 2.02649819 2.0264982
#>  [7,] 0.057676880 0.50885957 1.2230948
#>  [8,] 2.176083449 0.66675402 3.1480825
#>  [9,] 1.585278410 0.51363421 3.1578782
#> [10,] 0.686601856 0.12542049 0.6866019
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17174084 0.17174084 0.17174084
#>  [2,] 1.41267730 1.41267730 1.41267730
#>  [3,] 0.53744562 0.53744562 0.53744562
#>  [4,] 0.63781825 0.63781825 0.63781825
#>  [5,] 0.03303037 0.03303037 0.03303037
#>  [6,] 1.63776137 1.63776137 1.63776137
#>  [7,] 0.72498855 0.72498855 0.72498855
#>  [8,] 1.94609335 1.94609335 1.94609335
#>  [9,] 1.01155691 1.01155691 1.01155691
#> [10,] 1.34990370 1.34990370 1.34990370
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.19426664 0.48728979 2.0359629
#>  [2,] 0.97767375 0.56923453 0.5183001
#>  [3,] 4.07873938 0.16151405 4.1630190
#>  [4,] 2.32900105 2.89841985 2.1572099
#>  [5,] 0.24526802 0.52954429 2.1818512
#>  [6,] 0.94493538 0.04903869 0.4005607
#>  [7,] 1.06204362 0.88961055 1.4270622
#>  [8,] 0.63166547 0.96287000 1.2095756
#>  [9,] 0.08931787 0.95927295 0.7860415
#> [10,] 2.52706721 3.40398219 0.3844907

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.78380852 0.78380852 0.7838085
#>  [2,] 1.23187871 2.81894973 0.1352558
#>  [3,] 0.02340904 1.18654599 1.0693660
#>  [4,] 2.75126712 2.75126712 1.5174609
#>  [5,] 0.86532821 0.86532821 0.8653282
#>  [6,] 0.88799585 0.88799585 0.8879958
#>  [7,] 1.29756383 0.05163856 0.8714286
#>  [8,] 0.45922349 0.52587939 0.2473372
#>  [9,] 1.61552695 0.74628761 4.1439271
#> [10,] 2.01738038 0.86763286 2.0173804
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.10336866 2.10336866 2.10336866
#>  [2,] 4.27068501 4.27068501 4.27068501
#>  [3,] 0.60021826 0.60021826 0.60021826
#>  [4,] 1.05922133 1.05922133 1.05922133
#>  [5,] 0.04090319 0.04090319 0.04090319
#>  [6,] 0.08065267 0.08065267 0.08065267
#>  [7,] 1.12378871 1.12378871 1.12378871
#>  [8,] 0.61745721 0.61745721 0.61745721
#>  [9,] 3.84222830 3.84222830 3.84222830
#> [10,] 0.22820898 0.22820898 0.22820898
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.07952598 0.3511914 0.11465734
#>  [2,] 1.79706875 2.7580103 0.13525881
#>  [3,] 1.26298041 1.9445920 0.16943249
#>  [4,] 0.59160042 4.2539845 0.51104257
#>  [5,] 2.50082333 1.2920268 0.05978613
#>  [6,] 0.23619963 3.1961880 0.35783302
#>  [7,] 0.61631941 1.1617187 0.53211966
#>  [8,] 1.75235529 0.7374115 1.27718917
#>  [9,] 0.97166770 1.0620504 3.54842936
#> [10,] 1.83122250 0.5560359 0.55435420

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1630731  0.1265178 1.2219118
#>  [2,] 0.4276890  0.9662035 2.4757562
#>  [3,] 0.5480896  1.1355872 1.3311310
#>  [4,] 4.6163127  2.0558851 3.5787783
#>  [5,] 0.2343568  1.0385363 3.9475944
#>  [6,] 2.3745665  0.0456188 2.3745665
#>  [7,] 0.1963645 11.8050133 0.5365514
#>  [8,] 0.5387628  0.7235726 0.7461969
#>  [9,] 5.6742803  2.1869513 2.5068852
#> [10,] 3.1745102  1.4579524 1.0821224
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3600735 1.3600735 1.3600735
#>  [2,] 1.0900588 1.0900588 1.0900588
#>  [3,] 0.7523643 0.7523643 0.7523643
#>  [4,] 1.7163791 1.7163791 1.7163791
#>  [5,] 0.6805289 0.6805289 0.6805289
#>  [6,] 0.5351759 0.5351759 0.5351759
#>  [7,] 0.8066805 0.8066805 0.8066805
#>  [8,] 2.3964684 2.3964684 2.3964684
#>  [9,] 0.3892615 0.3892615 0.3892615
#> [10,] 0.5758032 0.5758032 0.5758032
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.68327379 0.65474716 0.18092888
#>  [2,] 0.16905424 0.53872949 0.11060166
#>  [3,] 0.31789021 1.54233706 0.61078865
#>  [4,] 0.09045357 0.27075886 0.26420481
#>  [5,] 1.47475957 4.03190594 0.03200984
#>  [6,] 3.88773866 0.43411555 0.30435458
#>  [7,] 1.28107373 1.27783807 2.40302295
#>  [8,] 0.09614133 0.24818137 0.26331711
#>  [9,] 0.48995567 0.91781002 0.15981789
#> [10,] 2.70480153 0.03849057 1.11186467

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3847473 0.8377081 0.1460304
#>  [2,] 0.5244960 1.4506362 2.5669175
#>  [3,] 1.3887597 1.3887597 0.1892509
#>  [4,] 1.7467846 0.5005982 4.4192094
#>  [5,] 1.3405021 0.5837802 0.4090563
#>  [6,] 0.1391049 3.3741109 5.4693985
#>  [7,] 0.1070889 0.3393124 0.2319427
#>  [8,] 1.0336202 1.0336202 1.0336202
#>  [9,] 0.1073275 1.5636173 0.1056837
#> [10,] 1.8703577 3.4166894 3.5320719
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3052683 0.3052683 0.3052683
#>  [2,] 0.2111989 0.2111989 0.2111989
#>  [3,] 2.4409998 2.4409998 2.4409998
#>  [4,] 1.5260728 1.5260728 1.5260728
#>  [5,] 0.8183596 0.8183596 0.8183596
#>  [6,] 2.5228505 2.5228505 2.5228505
#>  [7,] 0.4941066 0.4941066 0.4941066
#>  [8,] 0.3384198 0.3384198 0.3384198
#>  [9,] 3.0932938 3.0932938 3.0932938
#> [10,] 0.1146478 0.1146478 0.1146478
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.4616476 0.9244050 2.17202197
#>  [2,] 1.9007594 4.3785409 2.55390643
#>  [3,] 0.4491778 1.6370250 1.91656949
#>  [4,] 1.4289353 0.0214949 0.15076998
#>  [5,] 0.2210938 3.4700046 0.13708758
#>  [6,] 1.2203121 0.7959279 1.71563080
#>  [7,] 0.3985540 1.1026468 1.44252604
#>  [8,] 0.2819360 0.1312577 0.01557535
#>  [9,] 0.5072859 1.8205330 0.59294079
#> [10,] 0.8797461 1.0454779 0.12833588

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.47054334 1.30765335 0.4350021
#>  [2,] 0.85756094 0.85756094 0.8575609
#>  [3,] 0.17627025 0.76519083 0.5941254
#>  [4,] 1.08798055 1.36501741 3.3393833
#>  [5,] 1.28880016 1.28880016 1.2888002
#>  [6,] 1.04791215 0.06990744 0.7783910
#>  [7,] 0.19184442 1.79323233 0.7717551
#>  [8,] 0.02449578 0.64064086 1.1759905
#>  [9,] 0.75216243 0.75216243 0.7521624
#> [10,] 0.22746664 0.10742027 0.1074203
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08814549 0.56525166 0.08814549
#>  [2,] 0.37588160 0.53173003 0.53173003
#>  [3,] 0.15572804 0.78149568 0.55972605
#>  [4,] 0.72731852 1.37212612 0.50020409
#>  [5,] 0.18683070 0.74755665 0.71896898
#>  [6,] 0.51593372 0.22845625 0.46144087
#>  [7,] 2.00805402 0.48407650 0.14941294
#>  [8,] 0.24743100 0.08859365 0.73933316
#>  [9,] 0.35523532 0.95834729 0.07698219
#> [10,] 0.63670912 0.64010423 1.58227450
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.7763215 0.01253765 0.01253765
#>  [2,] 0.3514184 0.35141839 0.35141839
#>  [3,] 0.1971625 1.12432166 0.03857498
#>  [4,] 0.5703413 0.38503365 0.57034126
#>  [5,] 2.0511637 2.05116374 0.31430784
#>  [6,] 0.3796489 3.60914307 0.28567564
#>  [7,] 2.7012237 0.06395893 0.10300760
#>  [8,] 0.4346358 0.48376027 0.43463581
#>  [9,] 0.4287954 0.42879540 0.42879540
#> [10,] 0.3329344 0.33293437 0.74074710
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33516273 1.68491896 1.19459886
#>  [2,] 0.40458546 0.08930909 0.32215385
#>  [3,] 0.28170104 0.28170104 1.63467967
#>  [4,] 0.50026099 0.50026099 0.23803769
#>  [5,] 1.16407297 0.41018139 1.19114462
#>  [6,] 0.05880251 0.67599027 0.67599027
#>  [7,] 0.26538807 0.06350663 1.06070043
#>  [8,] 0.13088425 0.13088425 0.36074004
#>  [9,] 0.93514378 0.25689370 0.32353177
#> [10,] 0.09778143 0.09778143 0.09778143
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.04982248 0.47579713 0.4757971
#>  [2,] 0.89804261 0.04200514 1.9005758
#>  [3,] 0.06128016 2.57011102 0.4361573
#>  [4,] 0.38921911 0.38921911 0.2793358
#>  [5,] 0.94350693 0.94350693 0.9435069
#>  [6,] 1.12670258 2.77947866 1.4939830
#>  [7,] 3.16210482 0.36570137 1.9976008
#>  [8,] 0.69111506 0.69111506 0.6966021
#>  [9,] 0.56593323 1.12126830 0.0333685
#> [10,] 0.69118975 0.72229936 0.3305244

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.90886267 0.7216148 0.90886267
#>  [2,] 1.15183565 0.2752103 1.68781273
#>  [3,] 0.06038004 0.9834724 0.98347238
#>  [4,] 0.56382038 1.0823616 1.08236156
#>  [5,] 1.27883691 3.2226131 0.60443072
#>  [6,] 0.60242246 2.1920424 0.09876635
#>  [7,] 0.18233465 0.6319263 0.20506223
#>  [8,] 0.21983645 0.2198365 0.21983645
#>  [9,] 1.04994810 1.0499481 0.17755306
#> [10,] 1.27668471 1.4487781 0.04762013
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12891415 0.58877146 0.58877146
#>  [2,] 3.08102904 0.33561428 0.73452066
#>  [3,] 0.01890087 0.01890087 0.01890087
#>  [4,] 3.60894265 0.75739585 1.19636726
#>  [5,] 0.08534455 0.93952400 0.24182361
#>  [6,] 1.67046356 0.49851699 0.51116216
#>  [7,] 3.23442075 3.59758451 1.30414152
#>  [8,] 0.63369924 0.57568997 0.09059407
#>  [9,] 0.15568563 1.81873600 0.56084378
#> [10,] 0.15957712 5.08629401 3.88617431
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03398254 1.22598202 0.37530629
#>  [2,] 1.31091235 0.47890497 1.18447575
#>  [3,] 0.26529398 1.20704451 1.36094180
#>  [4,] 0.23508622 1.10753713 1.10753713
#>  [5,] 0.84283444 1.91789774 2.50706127
#>  [6,] 1.39271457 1.39271457 1.39271457
#>  [7,] 2.05477083 0.08886894 2.33403559
#>  [8,] 2.89641395 1.32272394 2.89641395
#>  [9,] 1.53849379 0.22316171 1.53849379
#> [10,] 0.09616046 0.09616046 0.09616046
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.05132608 1.67197708 1.34136854
#>  [2,] 0.02383869 0.02383869 0.02383869
#>  [3,] 0.38019921 1.29364215 4.43626181
#>  [4,] 0.68250685 0.68250685 0.68250685
#>  [5,] 0.03315066 0.91462698 2.20724224
#>  [6,] 0.08330312 1.47660708 0.89268237
#>  [7,] 0.64256738 0.08402627 0.79345546
#>  [8,] 0.93561767 1.22498727 1.22498727
#>  [9,] 2.21100485 5.58046658 0.86612380
#> [10,] 0.03001265 1.21961509 2.40441890
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.08040885 0.3660953 3.04336413
#>  [2,] 1.25685454 2.3208251 0.00487944
#>  [3,] 0.48189219 1.7475417 1.74754171
#>  [4,] 0.79334954 1.7052421 0.92795094
#>  [5,] 4.35602681 0.1472177 1.23861178
#>  [6,] 0.56856319 1.6028965 1.66917284
#>  [7,] 0.96345599 0.9634560 0.96345599
#>  [8,] 0.46985997 0.3543005 0.16667932
#>  [9,] 2.37911499 0.2462949 0.05260019
#> [10,] 0.72582941 1.6875218 0.34171794

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.09537901 1.09537901 0.25551243
#>  [2,] 0.26821211 0.26821211 0.26821211
#>  [3,] 1.20940379 1.63432180 1.76396993
#>  [4,] 0.00328976 0.00328976 0.00328976
#>  [5,] 0.17753849 0.17753849 0.17753849
#>  [6,] 0.27966222 0.27966222 0.27966222
#>  [7,] 0.55786447 0.55786447 0.55786447
#>  [8,] 0.94508044 0.23568125 0.23568125
#>  [9,] 0.97631593 0.12450287 0.97631593
#> [10,] 0.15867706 0.43291189 0.06995378
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.1159531 0.6270092 0.09541515
#>  [2,] 0.3679490 0.3679490 0.36794899
#>  [3,] 0.9029813 0.1790047 0.17900475
#>  [4,] 0.1209686 0.1209686 0.12096863
#>  [5,] 1.3218533 2.3227547 1.05853873
#>  [6,] 0.6174663 0.6174663 0.61746626
#>  [7,] 1.0007467 1.1829162 0.45637745
#>  [8,] 0.2048645 0.2048645 0.20486451
#>  [9,] 0.9705969 0.6592299 0.65522076
#> [10,] 0.5286707 0.5286707 1.65992358
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08396428 0.08396428 0.08396428
#>  [2,] 0.15007225 0.15007225 0.15007225
#>  [3,] 2.05499807 0.17939555 2.73265419
#>  [4,] 0.11429088 1.27082248 0.17548951
#>  [5,] 0.24668779 0.24668779 0.05661739
#>  [6,] 0.07514766 1.27244901 1.27244901
#>  [7,] 2.07477219 1.25042928 1.31245824
#>  [8,] 0.13017542 0.13017542 0.13017542
#>  [9,] 1.17744952 1.17744952 1.17744952
#> [10,] 0.30064317 0.30064317 0.11708595
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21706991 0.21706991 0.21706991
#>  [2,] 0.14551699 0.14551699 0.14551699
#>  [3,] 0.08905425 0.08905425 0.08905425
#>  [4,] 0.02866483 0.66156774 0.94069591
#>  [5,] 0.23667873 1.69975929 1.24770365
#>  [6,] 0.16861580 1.42630188 0.84110473
#>  [7,] 0.50865466 0.93994380 0.93994380
#>  [8,] 0.18936833 1.00574067 0.18936833
#>  [9,] 0.81770264 0.19809997 0.19809997
#> [10,] 0.62529608 0.36945306 0.44797904
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5795661 0.5795661 0.5094147
#>  [2,] 0.2337213 0.2337213 0.2337213
#>  [3,] 0.2270884 0.6235107 0.4427635
#>  [4,] 0.2634579 1.4234202 0.2634579
#>  [5,] 1.4655402 0.1578948 1.4655402
#>  [6,] 0.8999328 0.8999328 0.3537714
#>  [7,] 1.2055680 0.3824481 2.5044954
#>  [8,] 0.3270437 0.3246197 0.5008106
#>  [9,] 0.5468581 0.5413279 2.1794793
#> [10,] 0.2843144 0.2843144 0.2843144

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20517236 0.20517236 0.22603516
#>  [2,] 0.19253709 0.19253709 0.05936595
#>  [3,] 0.62347548 0.78587644 0.62433810
#>  [4,] 0.24263367 0.15082083 0.08103000
#>  [5,] 0.21552665 0.73571005 0.58301988
#>  [6,] 0.24343691 0.49615468 0.49615468
#>  [7,] 0.03239939 0.03239939 0.26313630
#>  [8,] 0.31873082 0.35322335 1.01702640
#>  [9,] 1.31700464 1.46072021 0.41066571
#> [10,] 0.61375200 0.67194507 0.11934552
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31781310 0.31781310 0.08425765
#>  [2,] 0.20585150 0.35351412 0.35351412
#>  [3,] 0.00821659 0.30754993 0.00821659
#>  [4,] 0.29911970 0.29911970 0.52895357
#>  [5,] 0.11229941 0.89738561 0.39350075
#>  [6,] 0.30125665 0.20466589 0.61849500
#>  [7,] 0.15866421 0.01825217 0.50049872
#>  [8,] 0.85685642 0.57391722 1.57481504
#>  [9,] 0.61161178 0.61161178 0.61161178
#> [10,] 0.31871878 1.15006886 0.40369834
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12685539 0.12685539 0.12685539
#>  [2,] 0.16515123 0.65736785 0.08227792
#>  [3,] 0.17140693 0.10379490 0.15422295
#>  [4,] 0.46229368 0.46229368 0.14561042
#>  [5,] 0.13292006 0.10020737 0.10020737
#>  [6,] 0.08891535 0.18699345 0.47771334
#>  [7,] 0.06276180 0.33540840 0.33540840
#>  [8,] 0.45965391 0.12241198 0.95964116
#>  [9,] 0.69636727 0.69636727 0.09321919
#> [10,] 0.43343925 0.01136628 0.01136628
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08489379 0.08489379 0.08489379
#>  [2,] 0.82565506 0.07139714 0.14176546
#>  [3,] 0.40959595 0.22810295 0.73702272
#>  [4,] 0.78640121 0.03443792 0.89654564
#>  [5,] 0.04713903 0.04713903 0.04713903
#>  [6,] 0.39561372 0.82182341 0.11276381
#>  [7,] 0.41874399 0.22612605 0.22612605
#>  [8,] 0.58480086 0.26713973 0.58480086
#>  [9,] 0.08527597 0.08527597 0.08527597
#> [10,] 0.69230844 0.40017332 0.11752938

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.55841809 0.19428402 0.51787290
#>  [2,] 0.25540352 0.69909341 0.39375014
#>  [3,] 0.06120656 0.04846774 0.06120656
#>  [4,] 0.45166835 0.16998463 1.18917167
#>  [5,] 0.28737029 0.52076571 1.39391791
#>  [6,] 0.90660778 0.34935317 0.73732514
#>  [7,] 0.58477751 0.26249934 0.05572198
#>  [8,] 0.19488227 0.23910518 0.23910518
#>  [9,] 0.23540511 0.51423536 0.19798493
#> [10,] 0.77566239 0.44127888 0.10940546
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.53150181 0.01839436 0.01839436
#>  [2,] 1.58056294 0.44915312 0.26013333
#>  [3,] 0.35856585 0.35523158 0.75373795
#>  [4,] 0.23137034 0.01341202 0.01341202
#>  [5,] 0.33537934 0.25829926 0.33537934
#>  [6,] 0.10403140 0.10403140 0.10403140
#>  [7,] 0.26049539 0.01091601 0.44720054
#>  [8,] 0.38434768 0.38434768 0.37507558
#>  [9,] 0.02963708 0.02963708 0.02963708
#> [10,] 0.04888399 0.08069123 0.08069123
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20566273 0.24446953 0.49160882
#>  [2,] 0.44133245 0.29728802 0.10875385
#>  [3,] 0.24408387 0.06347356 0.14201019
#>  [4,] 0.66100263 0.05664243 0.13921738
#>  [5,] 0.02683884 0.02683884 0.02683884
#>  [6,] 0.85357208 0.52438589 0.08156545
#>  [7,] 0.08974338 0.08974338 0.08974338
#>  [8,] 0.13467418 0.13467418 0.13467418
#>  [9,] 0.44598501 1.31932389 0.38966638
#> [10,] 0.46296767 0.11290888 0.60100361
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02185838 0.02185838 0.02185838
#>  [2,] 0.35953071 0.06394658 0.20495911
#>  [3,] 0.13705833 0.13705833 0.13705833
#>  [4,] 0.54022845 0.72304572 1.24643993
#>  [5,] 0.10268708 0.10445410 0.10268708
#>  [6,] 0.11332738 0.32287732 0.15915581
#>  [7,] 0.53606485 0.17007419 0.32710119
#>  [8,] 0.06396835 0.60121281 0.24773587
#>  [9,] 0.11111843 0.10253809 0.11111843
#> [10,] 0.45880503 0.45880503 0.70095132

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.02049553 0.02049553 0.053242514
#>  [2,] 1.19618016 1.08170543 0.008039848
#>  [3,] 0.62001456 0.24788268 0.247882682
#>  [4,] 0.07560790 0.07609350 0.075607897
#>  [5,] 0.51078427 0.02398107 0.023981075
#>  [6,] 0.31549651 0.31549651 0.315496511
#>  [7,] 0.04044384 0.04044384 0.040443836
#>  [8,] 0.12598621 0.08688821 0.086888210
#>  [9,] 0.02144956 0.02144956 0.021449559
#> [10,] 0.64610808 0.71832754 0.624052696
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.37779633 0.37779633 0.37779633
#>  [2,] 0.74120912 0.18273109 0.58306360
#>  [3,] 0.05689239 0.28176431 0.31316714
#>  [4,] 0.05397237 0.05397237 0.05397237
#>  [5,] 0.09373487 0.09373487 0.09373487
#>  [6,] 0.32412785 0.07737361 0.32412785
#>  [7,] 0.02667092 0.22562376 0.02667092
#>  [8,] 0.51442458 0.35378161 0.22482116
#>  [9,] 0.55748431 0.70115826 0.55748431
#> [10,] 0.20426495 0.05391844 0.05391844
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04948551 0.04948551 0.04948551
#>  [2,] 0.63334765 0.71919072 0.50572938
#>  [3,] 0.13026807 0.13026807 0.58103854
#>  [4,] 0.66611193 0.37590992 0.72358666
#>  [5,] 0.64836331 0.64836331 0.64836331
#>  [6,] 0.03533186 0.70008056 0.36809491
#>  [7,] 0.17823496 0.12745103 0.96964906
#>  [8,] 0.20868269 0.20868269 0.20868269
#>  [9,] 0.04090186 1.20231929 0.43071220
#> [10,] 0.50022374 0.21287037 0.07055094
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.11084376 0.085069307 0.11084376
#>  [2,] 0.01588323 0.015883229 0.01588323
#>  [3,] 0.08320131 0.537739919 0.05736220
#>  [4,] 0.59199494 0.065388185 0.15835033
#>  [5,] 0.38782695 0.260605128 0.26060513
#>  [6,] 1.14139138 1.263249554 0.10523812
#>  [7,] 0.23711781 0.100685173 0.57663355
#>  [8,] 0.43823908 0.438239079 0.18806054
#>  [9,] 0.47847060 1.224688533 0.20391588
#> [10,] 0.54772549 0.001960208 0.54772549