Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.21025340 0.13701421 0.76145587
#>  [2,] 1.06021229 2.23612715 2.23612715
#>  [3,] 0.12694403 0.03448185 0.12694403
#>  [4,] 0.70725605 0.75796762 4.33137550
#>  [5,] 0.06343374 0.06343374 0.06343374
#>  [6,] 1.05313186 3.62379030 1.30278288
#>  [7,] 6.91703926 0.09961028 1.60909189
#>  [8,] 0.86499519 0.48405242 0.24733575
#>  [9,] 2.19407877 0.63584302 4.29825823
#> [10,] 3.15026370 0.01231094 0.61062541
rpextmo(10, 3, a = 1)      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.7633164 1.7633164 1.7633164
#>  [2,] 1.1122071 1.1122071 1.1122071
#>  [3,] 0.9682381 0.9682381 0.9682381
#>  [4,] 0.3457255 0.3457255 0.3457255
#>  [5,] 0.3741206 0.3741206 0.3741206
#>  [6,] 8.0135594 8.0135594 8.0135594
#>  [7,] 1.1185018 1.1185018 1.1185018
#>  [8,] 2.7365318 2.7365318 2.7365318
#>  [9,] 0.4314551 0.4314551 0.4314551
#> [10,] 3.5259471 3.5259471 3.5259471
rpextmo(10, 3, b = 1)      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 2.3221298 0.93736200 1.0647741
#>  [2,] 0.7520642 0.08541561 2.1834367
#>  [3,] 3.0514465 1.33114984 0.2022977
#>  [4,] 1.0819536 0.24721281 0.9559147
#>  [5,] 0.2041259 0.94403866 0.0550832
#>  [6,] 0.9524135 2.26837695 1.1221962
#>  [7,] 0.9055229 0.77126181 2.1551276
#>  [8,] 1.4529971 0.96994658 0.7247798
#>  [9,] 0.3488292 1.29158281 0.2457238
#> [10,] 0.5706252 1.30688172 0.4307214

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.0953067 1.4158864 0.03413508
#>  [2,] 0.7190224 1.0230549 1.75004829
#>  [3,] 0.2108896 2.3227331 0.69123424
#>  [4,] 0.4408353 0.1150108 0.44083534
#>  [5,] 4.0511928 1.4015304 0.63707891
#>  [6,] 1.5994054 1.0401886 1.33078219
#>  [7,] 5.0465284 0.2516251 8.01931689
#>  [8,] 0.6435314 0.6435314 0.64353135
#>  [9,] 1.4881440 1.4881440 0.81862006
#> [10,] 0.9946290 0.1526332 1.92763462
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.68464296 1.68464296 1.68464296
#>  [2,] 2.41433011 2.41433011 2.41433011
#>  [3,] 0.81961203 0.81961203 0.81961203
#>  [4,] 0.98791753 0.98791753 0.98791753
#>  [5,] 1.25781999 1.25781999 1.25781999
#>  [6,] 1.67524859 1.67524859 1.67524859
#>  [7,] 0.32208252 0.32208252 0.32208252
#>  [8,] 0.78067774 0.78067774 0.78067774
#>  [9,] 0.66223995 0.66223995 0.66223995
#> [10,] 0.01789456 0.01789456 0.01789456
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.57849588 0.8742561 0.18695697
#>  [2,] 2.64869557 0.7750317 0.67447098
#>  [3,] 0.77813321 2.2268747 0.40093761
#>  [4,] 0.85250111 0.4443317 1.16210105
#>  [5,] 0.37118793 0.3363506 0.02028462
#>  [6,] 0.03242256 3.7137989 0.24756012
#>  [7,] 1.50285544 1.8766934 1.02390162
#>  [8,] 0.06092447 0.7795975 0.93242603
#>  [9,] 0.88018017 0.9401217 0.92395114
#> [10,] 0.18399935 1.1291580 0.77861511

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5981607 1.3499544 2.47857310
#>  [2,] 0.5820778 1.1060661 2.10523573
#>  [3,] 0.9068701 0.9068701 0.05702708
#>  [4,] 0.0651952 1.0678556 0.70591453
#>  [5,] 0.5867142 3.1259864 2.03856764
#>  [6,] 0.3287700 4.4574813 3.59399445
#>  [7,] 1.6961425 1.6961425 0.81101203
#>  [8,] 1.1009816 1.1009816 0.69021791
#>  [9,] 1.1795894 0.4288042 1.08692485
#> [10,] 1.4114583 1.3953767 2.25152268
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4956983 0.4956983 0.4956983
#>  [2,] 0.6763519 0.6763519 0.6763519
#>  [3,] 1.7103028 1.7103028 1.7103028
#>  [4,] 0.5179200 0.5179200 0.5179200
#>  [5,] 0.5037067 0.5037067 0.5037067
#>  [6,] 0.5427168 0.5427168 0.5427168
#>  [7,] 0.0933435 0.0933435 0.0933435
#>  [8,] 0.1855984 0.1855984 0.1855984
#>  [9,] 1.6182660 1.6182660 1.6182660
#> [10,] 0.7694942 0.7694942 0.7694942
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9076699 0.75392995 0.9165754
#>  [2,] 0.2120281 2.56506094 1.0515223
#>  [3,] 0.8786827 1.61074777 2.5900330
#>  [4,] 0.2344387 5.51706313 1.8008478
#>  [5,] 0.6748630 0.10087886 0.1816315
#>  [6,] 1.4371583 3.37501366 1.6903823
#>  [7,] 0.0370432 0.04926691 0.3568153
#>  [8,] 1.2472071 1.57336709 0.5735965
#>  [9,] 0.3312571 0.07684283 5.3101948
#> [10,] 0.4326051 0.73493724 0.1023019

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09082556 0.09082556 0.09082556
#>  [2,] 2.39239981 0.45103816 3.46059064
#>  [3,] 1.03220075 0.51177036 0.16890475
#>  [4,] 0.56690998 0.76231072 1.77280241
#>  [5,] 1.21316240 1.88706364 0.57487394
#>  [6,] 0.17916647 0.17916647 0.17916647
#>  [7,] 0.70085086 0.97988867 1.02127712
#>  [8,] 4.62209445 1.36745007 0.46241774
#>  [9,] 0.59186315 0.08311949 0.70208264
#> [10,] 2.31158119 1.81386929 0.14900510
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.841492149 0.841492149 0.841492149
#>  [2,] 4.133852755 4.133852755 4.133852755
#>  [3,] 0.008142716 0.008142716 0.008142716
#>  [4,] 1.558642488 1.558642488 1.558642488
#>  [5,] 0.682087231 0.682087231 0.682087231
#>  [6,] 0.035165050 0.035165050 0.035165050
#>  [7,] 0.444800704 0.444800704 0.444800704
#>  [8,] 1.784849045 1.784849045 1.784849045
#>  [9,] 0.361541585 0.361541585 0.361541585
#> [10,] 2.025178470 2.025178470 2.025178470
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 2.53332913 0.30091027 1.3926617
#>  [2,] 1.15474918 1.70634121 0.4524667
#>  [3,] 5.57996890 1.97021680 1.1757254
#>  [4,] 0.12704824 0.09682636 0.4692538
#>  [5,] 0.09116183 3.18969343 1.7778046
#>  [6,] 0.18143540 0.04421977 1.2577181
#>  [7,] 3.01635110 0.61257081 0.6696754
#>  [8,] 0.48145787 0.89152899 0.7647680
#>  [9,] 0.39727530 0.08582515 1.6060668
#> [10,] 0.91111745 1.81931321 0.3879386

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2329652 1.0081029 1.10120121
#>  [2,] 0.4210062 1.7395735 3.57420566
#>  [3,] 0.5579768 2.1893589 0.06018503
#>  [4,] 1.3059557 1.6329572 2.42305469
#>  [5,] 0.6778933 0.6778933 0.67789327
#>  [6,] 0.7918366 0.5480217 0.28936112
#>  [7,] 1.1772013 1.9607413 1.39925185
#>  [8,] 1.4760671 1.4152546 0.51296803
#>  [9,] 1.1843841 1.1843841 1.18438408
#> [10,] 5.2056316 1.4114701 0.79343697
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5248014 0.5248014 0.5248014
#>  [2,] 0.5334518 0.5334518 0.5334518
#>  [3,] 1.1458581 1.1458581 1.1458581
#>  [4,] 0.5828912 0.5828912 0.5828912
#>  [5,] 0.3160227 0.3160227 0.3160227
#>  [6,] 2.0604277 2.0604277 2.0604277
#>  [7,] 0.4459985 0.4459985 0.4459985
#>  [8,] 1.3614103 1.3614103 1.3614103
#>  [9,] 0.6148185 0.6148185 0.6148185
#> [10,] 1.3352898 1.3352898 1.3352898
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.01588931 1.53907869 0.7754173
#>  [2,] 0.29261826 0.98353090 1.0042487
#>  [3,] 1.28283067 0.04240469 0.2829267
#>  [4,] 1.93335129 0.84560655 0.5525455
#>  [5,] 1.14413590 0.83778530 0.5814147
#>  [6,] 1.70618492 0.56879658 0.1034190
#>  [7,] 0.01790935 0.49672098 1.6085282
#>  [8,] 0.57428605 0.58604469 1.2119652
#>  [9,] 0.14424273 0.99391085 0.6427932
#> [10,] 0.40115287 0.12060227 1.0452904

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1848138 0.2118083 2.2976103
#>  [2,] 0.2341875 0.2341875 0.2341875
#>  [3,] 0.4954054 3.0795692 1.9899051
#>  [4,] 1.6632440 1.3234657 0.5598160
#>  [5,] 2.0724113 2.1264487 1.1165902
#>  [6,] 1.7353632 0.7026228 1.7353632
#>  [7,] 3.1457848 3.0198934 0.4738879
#>  [8,] 1.3057145 0.2874185 1.3352264
#>  [9,] 0.6652690 0.1247361 0.5217184
#> [10,] 1.4674900 0.3998523 3.2942847
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.66102134 0.66102134 0.66102134
#>  [2,] 0.22133556 0.22133556 0.22133556
#>  [3,] 0.02758008 0.02758008 0.02758008
#>  [4,] 1.65236977 1.65236977 1.65236977
#>  [5,] 1.32435693 1.32435693 1.32435693
#>  [6,] 0.62441056 0.62441056 0.62441056
#>  [7,] 1.86772580 1.86772580 1.86772580
#>  [8,] 1.91819987 1.91819987 1.91819987
#>  [9,] 1.62351791 1.62351791 1.62351791
#> [10,] 0.04062028 0.04062028 0.04062028
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.33268560 0.50586470 0.7606320
#>  [2,] 4.64646854 2.10224135 1.8362233
#>  [3,] 0.07638958 0.20456853 0.3828758
#>  [4,] 0.45526037 0.92277589 0.8275574
#>  [5,] 0.55684758 0.14269014 0.3995251
#>  [6,] 0.60091409 0.02603663 0.6160883
#>  [7,] 2.42715863 1.48808974 3.0603275
#>  [8,] 0.35273816 0.63694811 0.6099293
#>  [9,] 0.92620446 0.16397770 2.0031228
#> [10,] 0.75468557 0.14878365 0.2671162

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7345830 2.2132642 1.1344171
#>  [2,] 1.1552259 1.6678280 0.4328176
#>  [3,] 0.9767957 0.9767957 0.9767957
#>  [4,] 0.2581773 0.2581773 0.2581773
#>  [5,] 0.4276926 0.4276926 0.4276926
#>  [6,] 2.9028605 3.3777143 3.6066687
#>  [7,] 0.2784249 0.7592875 0.3710035
#>  [8,] 3.0280914 1.3592712 0.6058629
#>  [9,] 0.2610214 2.6632672 0.4811658
#> [10,] 0.2602782 0.1421737 0.9889961
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1194394 0.1194394 0.1194394
#>  [2,] 0.8098342 0.8098342 0.8098342
#>  [3,] 0.4041502 0.4041502 0.4041502
#>  [4,] 0.1162137 0.1162137 0.1162137
#>  [5,] 3.8159203 3.8159203 3.8159203
#>  [6,] 0.5896482 0.5896482 0.5896482
#>  [7,] 0.3418438 0.3418438 0.3418438
#>  [8,] 0.4147487 0.4147487 0.4147487
#>  [9,] 2.8557126 2.8557126 2.8557126
#> [10,] 2.0045229 2.0045229 2.0045229
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>            [,1]        [,2]       [,3]
#>  [1,] 0.5336553 1.612389360 0.13358857
#>  [2,] 1.4733234 0.280309809 1.08991049
#>  [3,] 1.4004140 0.009816909 1.01512927
#>  [4,] 0.8073097 0.087241424 2.03578433
#>  [5,] 1.7376920 0.121684288 0.85382085
#>  [6,] 0.8932862 0.071896864 1.55196431
#>  [7,] 0.3786367 1.007856086 0.94910061
#>  [8,] 1.0170423 1.695372611 0.83537216
#>  [9,] 1.5354584 0.042965939 0.01227236
#> [10,] 0.1629448 0.948563919 0.14381273

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1622423 1.0981833 0.4162885
#>  [2,] 0.5646863 1.7052032 1.7052032
#>  [3,] 3.0215187 2.7274929 0.1600081
#>  [4,] 0.6146511 0.6146511 0.3125167
#>  [5,] 1.7761165 4.3535114 2.5651214
#>  [6,] 0.5026095 0.1695773 4.7398973
#>  [7,] 0.4776642 1.4312253 4.5451669
#>  [8,] 0.1969008 0.5707688 0.2785757
#>  [9,] 1.6341495 1.3688553 0.7441691
#> [10,] 1.2967409 0.3077163 1.2967409
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3135413 0.3135413 0.3135413
#>  [2,] 3.0837919 3.0837919 3.0837919
#>  [3,] 1.0486008 1.0486008 1.0486008
#>  [4,] 0.2402651 0.2402651 0.2402651
#>  [5,] 0.2029432 0.2029432 0.2029432
#>  [6,] 1.5930838 1.5930838 1.5930838
#>  [7,] 3.2617076 3.2617076 3.2617076
#>  [8,] 0.8501977 0.8501977 0.8501977
#>  [9,] 0.5579231 0.5579231 0.5579231
#> [10,] 0.4069811 0.4069811 0.4069811
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.68681230 0.90594211 0.10002411
#>  [2,] 0.40077795 5.15537856 0.95716295
#>  [3,] 1.64339946 0.01573486 0.48566390
#>  [4,] 2.34024316 0.81530750 4.08929125
#>  [5,] 0.04578414 0.02145307 0.92046121
#>  [6,] 2.82346254 0.32494089 0.72610572
#>  [7,] 0.18732450 0.97758859 0.58126717
#>  [8,] 0.87922315 1.19137934 0.04380636
#>  [9,] 2.04488146 1.03530857 0.32031159
#> [10,] 1.35918128 0.25644603 0.04234567

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.7393126 1.1640199 1.91289271
#>  [2,] 0.5367133 0.4995146 2.04236542
#>  [3,] 0.2722404 0.2722404 0.27224044
#>  [4,] 0.7675156 1.1012593 3.41883077
#>  [5,] 0.6053955 1.3192749 2.68580414
#>  [6,] 1.0017491 1.6422679 0.26283035
#>  [7,] 0.4370649 2.2127786 1.17287788
#>  [8,] 0.4237984 0.8273174 0.21515789
#>  [9,] 0.7482057 0.7482057 0.09686184
#> [10,] 0.5650079 0.1533091 0.68749899
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.15210506 0.15210506 0.15210506
#>  [2,] 0.65167400 0.65167400 0.65167400
#>  [3,] 3.18227715 3.18227715 3.18227715
#>  [4,] 0.63141509 0.63141509 0.63141509
#>  [5,] 1.02209952 1.02209952 1.02209952
#>  [6,] 0.05596603 0.05596603 0.05596603
#>  [7,] 0.60822217 0.60822217 0.60822217
#>  [8,] 2.18129653 2.18129653 2.18129653
#>  [9,] 1.25063776 1.25063776 1.25063776
#> [10,] 0.71610279 0.71610279 0.71610279
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>              [,1]       [,2]       [,3]
#>  [1,] 2.102645169 0.24398594 1.25321432
#>  [2,] 0.892065492 2.94121234 2.02411948
#>  [3,] 0.652993084 0.04785638 0.56356112
#>  [4,] 1.111778829 0.08666474 1.41364345
#>  [5,] 1.577541757 0.91008663 0.06835123
#>  [6,] 0.311528275 0.82559198 1.08983892
#>  [7,] 0.071871949 0.31054379 0.56875861
#>  [8,] 0.004040327 0.89858090 0.70668266
#>  [9,] 0.549326987 0.79713408 0.72857161
#> [10,] 0.193051564 0.18198285 0.38034257

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.03327041 2.6074998 2.3260134
#>  [2,] 0.22586069 0.2258607 0.2258607
#>  [3,] 0.73062892 0.8166354 0.8166354
#>  [4,] 1.78835703 0.2973253 0.4808730
#>  [5,] 0.97431714 0.4757812 1.8557066
#>  [6,] 1.28467809 3.0945476 1.8814978
#>  [7,] 1.32458310 1.3245831 0.8621457
#>  [8,] 1.12325567 4.3035812 1.7962592
#>  [9,] 0.30976464 0.4086140 0.2576348
#> [10,] 1.01910357 0.3083199 2.2773711
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1162718 0.1162718 0.1162718
#>  [2,] 3.1563832 3.1563832 3.1563832
#>  [3,] 1.4476219 1.4476219 1.4476219
#>  [4,] 0.5563194 0.5563194 0.5563194
#>  [5,] 2.3871326 2.3871326 2.3871326
#>  [6,] 2.6820217 2.6820217 2.6820217
#>  [7,] 1.1956479 1.1956479 1.1956479
#>  [8,] 0.6057458 0.6057458 0.6057458
#>  [9,] 0.4006394 0.4006394 0.4006394
#> [10,] 1.9851009 1.9851009 1.9851009
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>              [,1]       [,2]       [,3]
#>  [1,] 0.012502515 0.80528447 0.03546989
#>  [2,] 0.516539448 1.63987179 0.03662454
#>  [3,] 4.184442962 0.44380327 0.38519626
#>  [4,] 0.011777712 0.46625061 0.28693541
#>  [5,] 0.755291176 0.05463573 0.15150582
#>  [6,] 0.404628124 0.51994430 0.60387615
#>  [7,] 0.316211452 0.41803172 2.54243912
#>  [8,] 0.933848994 1.30011521 0.81844609
#>  [9,] 0.003910411 0.77615259 0.83308050
#> [10,] 2.303063158 0.11151714 0.42909146

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.33813143 0.33813143 1.0671167
#>  [2,] 0.15297400 0.09661481 0.9252028
#>  [3,] 0.46548482 0.33375409 0.2297123
#>  [4,] 0.20918410 0.21727329 0.2172733
#>  [5,] 0.65431494 0.79094936 0.7909494
#>  [6,] 1.53703957 1.03954823 1.7252510
#>  [7,] 0.14775113 0.01206332 0.5707006
#>  [8,] 0.52268504 0.40417791 0.6041696
#>  [9,] 0.03846578 0.34204258 0.7618601
#> [10,] 0.57915208 0.40927859 2.3671226
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1335430 0.96038035 2.3194052
#>  [2,] 0.6440648 0.35134186 0.3513419
#>  [3,] 0.5468656 0.02933263 0.5468656
#>  [4,] 0.7253121 0.87925501 0.8792550
#>  [5,] 0.3297414 0.07619626 0.6044624
#>  [6,] 0.2009531 0.12859893 0.1285989
#>  [7,] 0.5178675 0.37864976 0.3786498
#>  [8,] 0.2217147 0.22171473 0.2217147
#>  [9,] 0.2470356 1.17862930 0.4463536
#> [10,] 0.3056098 0.13154247 0.3056098
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]        [,2]      [,3]
#>  [1,] 0.57693336 0.009069009 0.5769334
#>  [2,] 1.33035849 0.129352990 0.0395842
#>  [3,] 0.11099861 1.594574537 0.1109986
#>  [4,] 0.06352806 0.485218204 0.4489757
#>  [5,] 0.91144837 0.543361348 2.2853696
#>  [6,] 0.08627817 0.544234861 0.5442349
#>  [7,] 1.05424219 1.054242186 1.0542422
#>  [8,] 1.43687787 2.255032075 0.8974193
#>  [9,] 0.19985292 0.199852915 0.6393567
#> [10,] 0.25857027 0.258570270 0.2585703
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31986288 0.06749177 0.31986288
#>  [2,] 0.02352919 0.22122708 0.02352919
#>  [3,] 0.84939690 1.74037407 0.02067039
#>  [4,] 0.35110171 0.35110171 0.30186057
#>  [5,] 0.04700319 0.04700319 0.04700319
#>  [6,] 0.13447673 0.65060555 0.65060555
#>  [7,] 0.24168094 0.30471956 0.30471956
#>  [8,] 0.15226436 2.84254353 0.15226436
#>  [9,] 1.30948744 1.78134664 0.61771919
#> [10,] 1.34402817 0.44649476 1.54819377
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19590034 0.19590034 2.47114345
#>  [2,] 0.32121376 0.95249553 0.07710960
#>  [3,] 0.68889055 0.31863026 0.20979474
#>  [4,] 0.66699404 0.70385020 4.14156592
#>  [5,] 0.04349382 0.04349382 0.04349382
#>  [6,] 0.43945903 0.58210691 0.68777417
#>  [7,] 0.30628426 0.30628426 1.46444269
#>  [8,] 1.36978959 0.69347928 0.92868098
#>  [9,] 0.87418709 0.49673439 0.49673439
#> [10,] 0.27155094 0.27155094 1.31256808

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.27790112 0.12824944 0.37619110
#>  [2,] 2.45745925 2.45745925 2.45745925
#>  [3,] 1.90578587 0.35418589 1.35268878
#>  [4,] 0.04403973 0.04403973 0.04403973
#>  [5,] 2.36465267 1.67935728 0.81961925
#>  [6,] 1.53530026 0.89675977 0.89675977
#>  [7,] 0.20040923 0.20040923 0.20040923
#>  [8,] 0.04990045 0.04990045 0.04990045
#>  [9,] 1.27925918 0.97543358 1.37790959
#> [10,] 1.75832503 0.67124539 3.02662979
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.27178295 0.7385481 0.7385481
#>  [2,] 0.06584054 0.8405665 2.1366562
#>  [3,] 1.75052541 2.8531786 5.7144831
#>  [4,] 2.45680795 1.9525651 2.2880459
#>  [5,] 0.81368443 1.3999472 0.2710575
#>  [6,] 4.22772014 0.3991006 1.2597785
#>  [7,] 0.17364182 0.1189535 0.1189535
#>  [8,] 2.67976749 1.1829403 0.6161767
#>  [9,] 2.38027154 1.2011227 0.3588461
#> [10,] 2.45295582 0.4861868 2.4529558
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 3.92218379 0.61856502 0.26033084
#>  [2,] 0.84702868 1.89269283 5.74093079
#>  [3,] 2.24043836 1.99466548 2.80397209
#>  [4,] 1.83274384 0.32270860 0.53732321
#>  [5,] 0.09113593 0.09113593 0.09113593
#>  [6,] 0.59749642 0.63932715 2.17152780
#>  [7,] 0.56083758 3.54071498 1.68187874
#>  [8,] 1.33981914 1.32060364 1.28464347
#>  [9,] 0.30004818 2.80676804 7.84421716
#> [10,] 1.24311078 2.50309586 2.85841615
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.27847155 0.67920571 0.9643428
#>  [2,] 0.04085497 3.45158570 0.6484823
#>  [3,] 0.64016363 0.51516132 1.2513997
#>  [4,] 0.86450678 0.71728623 0.8645068
#>  [5,] 0.28166167 1.43636292 1.0133503
#>  [6,] 2.55302985 0.36604423 1.1461245
#>  [7,] 0.19618534 2.65944280 2.6543846
#>  [8,] 0.79226730 1.15398907 0.9198960
#>  [9,] 1.57850157 0.09366088 1.7145501
#> [10,] 0.77815104 1.65465790 1.9997880
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 3.1746281 5.3784587 6.1929253
#>  [2,] 1.5124652 0.8038051 0.8636334
#>  [3,] 6.2384320 2.1559457 0.4825365
#>  [4,] 1.5127186 0.5290901 1.5127186
#>  [5,] 1.1585938 1.0410540 0.7740538
#>  [6,] 1.6822108 2.7090055 4.6749690
#>  [7,] 1.2701447 1.2701447 0.6849750
#>  [8,] 0.5871955 0.5871955 0.5871955
#>  [9,] 1.8573100 4.9733610 3.5710729
#> [10,] 2.0622329 1.5907530 0.2374537

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.74866023 0.08332165 0.12363920
#>  [2,] 0.08302062 0.08302062 0.08302062
#>  [3,] 0.23494630 0.06218765 1.07186619
#>  [4,] 0.20274773 1.43165279 1.43165279
#>  [5,] 0.11954050 0.11954050 0.11954050
#>  [6,] 1.06416880 1.04655169 0.39501388
#>  [7,] 0.16159168 0.16159168 0.17374042
#>  [8,] 0.33782707 0.02191320 0.02191320
#>  [9,] 0.59551223 0.59551223 0.59551223
#> [10,] 0.10994915 0.10994915 0.01318763
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.02718271 0.39073774 0.91546151
#>  [2,] 0.02544486 0.02544486 0.02544486
#>  [3,] 0.98102556 0.98102556 0.13151899
#>  [4,] 0.40309129 0.40309129 0.30229554
#>  [5,] 1.48695238 1.03635240 0.55470050
#>  [6,] 0.80051483 0.80051483 0.57448701
#>  [7,] 0.95269500 0.64393080 0.14267527
#>  [8,] 0.95035542 0.49214710 2.31853118
#>  [9,] 0.01870795 0.01870795 0.01870795
#> [10,] 1.00408821 0.78572623 0.78572623
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42531914 0.15560264 0.22087043
#>  [2,] 0.79870294 0.79870294 0.19512593
#>  [3,] 0.70842507 0.70842507 0.70842507
#>  [4,] 1.23735526 0.75617555 1.70787645
#>  [5,] 1.12552189 1.07935182 1.08776962
#>  [6,] 0.36579648 0.26232333 0.36579648
#>  [7,] 0.06405411 0.06405411 0.06405411
#>  [8,] 0.09129742 0.09129742 0.09129742
#>  [9,] 0.88147409 0.88147409 0.88147409
#> [10,] 1.16196279 0.72582918 1.16196279
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46595648 0.23405748 0.23405748
#>  [2,] 0.12823710 0.10853641 0.12823710
#>  [3,] 0.16177602 0.06133158 0.16177602
#>  [4,] 0.09091994 0.37630696 0.37630696
#>  [5,] 0.30762436 0.30762436 0.30762436
#>  [6,] 1.00196802 1.00196802 0.07510111
#>  [7,] 1.86603545 1.41127525 0.12087431
#>  [8,] 0.33195460 1.53781292 0.33195460
#>  [9,] 0.65046569 0.65046569 0.45868779
#> [10,] 0.02786476 0.24053497 0.24053497
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84604012 0.03567440 0.84604012
#>  [2,] 0.24352048 1.64337815 2.23358898
#>  [3,] 0.21628898 0.21628898 0.21628898
#>  [4,] 0.00632537 0.00632537 0.00632537
#>  [5,] 0.28399208 0.28399208 0.28399208
#>  [6,] 0.11930954 0.11930954 1.07463266
#>  [7,] 1.39646470 0.69106963 0.83902904
#>  [8,] 1.67945378 0.46928794 1.57728744
#>  [9,] 0.43028989 0.43028989 0.43028989
#> [10,] 0.36594619 0.36594619 0.23593461

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.610889438 0.525616243 0.917541223
#>  [2,] 0.032781026 0.618174143 0.618174143
#>  [3,] 0.590111641 0.165128203 0.590111641
#>  [4,] 0.003157477 0.003157477 0.003157477
#>  [5,] 0.035799549 0.601016178 0.035799549
#>  [6,] 0.043409749 0.043409749 0.043409749
#>  [7,] 0.290613605 0.347604973 0.028748464
#>  [8,] 0.983361220 0.983361220 0.013916712
#>  [9,] 0.189064633 0.056093502 0.189064633
#> [10,] 0.078491174 0.078491174 0.078491174
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19501739 0.15701348 0.29050586
#>  [2,] 0.37521181 0.45157861 0.45157861
#>  [3,] 0.04223101 0.04223101 0.04223101
#>  [4,] 0.10301732 0.10301732 0.10301732
#>  [5,] 0.11328568 0.85742648 0.11328568
#>  [6,] 0.35715645 0.18532120 0.08178632
#>  [7,] 0.16676199 0.16676199 0.16676199
#>  [8,] 0.65129002 0.65129002 0.65129002
#>  [9,] 0.85348524 0.91683374 0.91683374
#> [10,] 0.39388250 0.07299423 0.39388250
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>               [,1]       [,2]       [,3]
#>  [1,] 0.0223769834 1.43258995 0.02237698
#>  [2,] 1.0034020575 0.63897307 0.72689752
#>  [3,] 0.2046234416 0.20462344 0.20462344
#>  [4,] 0.0970432258 0.19960941 0.19960941
#>  [5,] 0.8359907880 0.71297801 0.83599079
#>  [6,] 0.1594802792 0.04585915 0.06526916
#>  [7,] 0.0002817253 0.24844085 0.17313173
#>  [8,] 0.2598222805 0.18432372 0.38600807
#>  [9,] 0.3722713351 1.44309689 0.06999688
#> [10,] 0.2413711977 0.30783061 0.23443004
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1291615 1.12916154 0.78493869
#>  [2,] 0.1797875 0.17978748 0.17978748
#>  [3,] 1.2656399 0.67726559 1.13021565
#>  [4,] 0.3422125 0.43248818 0.43248818
#>  [5,] 0.2514269 0.10762565 0.35907159
#>  [6,] 0.0658138 0.82381011 0.82381011
#>  [7,] 0.2526894 1.00267293 0.06342967
#>  [8,] 0.9814463 0.09716661 0.58353500
#>  [9,] 0.6952013 0.10758210 0.08847233
#> [10,] 0.1789092 0.17890919 0.17890919

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.005913473 0.005913473 0.05596935
#>  [2,] 0.408154488 0.017268887 0.86510084
#>  [3,] 1.205580825 0.761925781 0.39822155
#>  [4,] 0.267056376 0.099810128 0.22707035
#>  [5,] 0.097806530 0.421401840 0.22838146
#>  [6,] 0.437384988 0.067697198 0.43738499
#>  [7,] 0.780716008 0.546642708 0.54664271
#>  [8,] 0.882661770 1.262422470 0.35860787
#>  [9,] 0.808454779 0.323060250 0.34121199
#> [10,] 0.774470202 1.114259184 0.58522572
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.04272230 0.25895408 0.0427223
#>  [2,] 0.12013229 0.37372879 0.9967904
#>  [3,] 0.02362509 1.58993690 0.1358346
#>  [4,] 0.10062640 0.49151839 0.9083276
#>  [5,] 0.00422840 0.00422840 0.0042284
#>  [6,] 0.16280733 0.16280733 0.1628073
#>  [7,] 0.05219482 0.67721944 0.2030210
#>  [8,] 0.44002674 0.42865685 0.4286569
#>  [9,] 0.23150762 0.08014262 1.2784393
#> [10,] 0.24280760 0.24280760 0.1298378
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]         [,2]       [,3]
#>  [1,] 0.21947319 0.1155686742 0.31311678
#>  [2,] 0.21167886 0.2116788595 0.24264222
#>  [3,] 0.20810281 0.2081028073 0.11001058
#>  [4,] 1.03443746 0.0966824317 0.37693265
#>  [5,] 0.33551274 0.3454892935 0.48847569
#>  [6,] 0.04169888 0.0856849330 0.01483085
#>  [7,] 0.13081458 0.2896063932 0.31898415
#>  [8,] 0.49733714 0.0007387831 0.22453080
#>  [9,] 0.25433148 0.2543314772 0.25433148
#> [10,] 0.90437351 0.8482804887 0.13980356
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03673222 1.06471749 0.18686377
#>  [2,] 0.47582274 0.05959928 0.47582274
#>  [3,] 0.93316522 0.11970183 0.41301904
#>  [4,] 0.01985250 0.78634719 0.78634719
#>  [5,] 0.31601346 0.22212954 0.31601346
#>  [6,] 0.11279672 0.11279672 0.04264013
#>  [7,] 0.52449653 0.21851583 0.31443418
#>  [8,] 0.11071336 0.11071336 0.43085985
#>  [9,] 1.29860936 0.12393688 0.28392730
#> [10,] 0.17657639 0.14651242 0.45687553

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>              [,1]        [,2]       [,3]
#>  [1,] 0.037467809 0.026961774 0.02696177
#>  [2,] 0.115700220 0.796014888 0.83627342
#>  [3,] 0.013684979 0.600790124 0.00586893
#>  [4,] 0.085306271 0.166296512 0.08530627
#>  [5,] 0.683819359 0.165254426 0.32432659
#>  [6,] 0.028994462 0.028994462 0.02899446
#>  [7,] 0.018622234 0.018622234 0.01862223
#>  [8,] 1.371111696 0.713971278 1.32639860
#>  [9,] 0.005242602 0.005242602 0.25266095
#> [10,] 0.084001224 0.140092746 0.48617044
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.41718160 1.417239067 0.309661782
#>  [2,] 0.18679788 0.186797882 0.186797882
#>  [3,] 0.12396124 0.055571118 0.014447343
#>  [4,] 0.02175145 0.021751449 0.021751449
#>  [5,] 0.49968653 0.805437528 0.805437528
#>  [6,] 0.95282711 0.218238592 0.718783051
#>  [7,] 0.21084575 0.210845752 0.210845752
#>  [8,] 0.13516514 0.004891907 0.004891907
#>  [9,] 0.84283633 0.919101751 0.196431301
#> [10,] 0.01401568 0.014015680 0.101756324
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.14321475 0.008528095 0.55149560
#>  [2,] 0.01290534 0.357569468 0.08030384
#>  [3,] 0.05118421 0.079056796 0.02944705
#>  [4,] 0.78950290 0.936707457 0.45856405
#>  [5,] 0.49900012 0.499000118 0.24484287
#>  [6,] 0.13204957 0.132049574 0.13204957
#>  [7,] 0.27749774 0.574364171 0.41712207
#>  [8,] 0.24069290 0.019964148 0.45255798
#>  [9,] 0.06991182 0.069911823 0.06991182
#> [10,] 0.04158456 0.029950889 0.02995089
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12265633 0.12265633 0.58394021
#>  [2,] 1.02052942 1.12175460 1.24649109
#>  [3,] 0.28753469 0.43380543 0.39632179
#>  [4,] 0.03435532 0.03435532 0.26926103
#>  [5,] 0.91264778 0.14583922 0.51535787
#>  [6,] 0.07657145 0.05336083 0.07657145
#>  [7,] 0.34304629 0.01616395 0.12845764
#>  [8,] 0.26940701 0.05752617 0.60331881
#>  [9,] 1.07801159 1.07801159 1.28127283
#> [10,] 0.16325221 0.24617928 0.45770916