Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > \beta t \}} + \alpha t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.113790345 0.113790345 0.113790345
#>  [2,] 0.175584510 0.175584510 0.175584510
#>  [3,] 2.338301074 2.612180021 2.612180021
#>  [4,] 0.629706301 0.629706301 0.028503030
#>  [5,] 0.002770766 0.002770766 0.002770766
#>  [6,] 0.677048509 0.260517984 1.509430846
#>  [7,] 0.425070358 0.710377057 2.298207490
#>  [8,] 1.809332494 0.037623605 2.522232244
#>  [9,] 1.432901467 5.586161924 4.146310512
#> [10,] 0.220477292 0.220477292 0.220477292
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46111176 0.46111176 0.46111176
#>  [2,] 0.37267328 0.37267328 0.37267328
#>  [3,] 0.05452417 0.05452417 0.05452417
#>  [4,] 0.20939187 0.20939187 0.20939187
#>  [5,] 0.47795125 0.47795125 0.47795125
#>  [6,] 3.33681762 3.33681762 3.33681762
#>  [7,] 0.45632701 0.45632701 0.45632701
#>  [8,] 0.22265734 0.22265734 0.22265734
#>  [9,] 0.64104695 0.64104695 0.64104695
#> [10,] 0.16712876 0.16712876 0.16712876
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.08839147 0.28985622 3.6974980
#>  [2,] 2.10282781 0.15930617 1.6919361
#>  [3,] 0.37798910 1.02411118 1.4519900
#>  [4,] 2.43483018 0.40999104 1.0110053
#>  [5,] 0.12024935 0.34961084 0.2565203
#>  [6,] 0.36302881 0.61518125 0.4693697
#>  [7,] 0.74919097 0.20448085 0.4205259
#>  [8,] 0.30942707 0.36291228 0.1735171
#>  [9,] 0.48511289 0.09907151 0.2557181
#> [10,] 0.76582003 5.48903965 0.2670437

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>              [,1]       [,2]      [,3]
#>  [1,] 1.060348039 0.79434253 1.6010392
#>  [2,] 0.479335068 0.47933507 0.4793351
#>  [3,] 0.020504712 0.43854937 0.2870272
#>  [4,] 0.859706509 0.01055783 1.1676743
#>  [5,] 0.501734856 0.50173486 0.5017349
#>  [6,] 0.847242576 0.62912370 1.1855416
#>  [7,] 0.115736173 0.11573617 0.1157362
#>  [8,] 0.007034003 2.38316469 4.0382813
#>  [9,] 2.187229583 2.44408754 1.4194137
#> [10,] 0.639459505 6.93628804 2.0476167
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.99007863 0.99007863 0.99007863
#>  [2,] 1.07373090 1.07373090 1.07373090
#>  [3,] 0.06502439 0.06502439 0.06502439
#>  [4,] 2.40054049 2.40054049 2.40054049
#>  [5,] 1.00831268 1.00831268 1.00831268
#>  [6,] 0.32897827 0.32897827 0.32897827
#>  [7,] 1.30898312 1.30898312 1.30898312
#>  [8,] 1.12063125 1.12063125 1.12063125
#>  [9,] 0.04665377 0.04665377 0.04665377
#> [10,] 0.30596076 0.30596076 0.30596076
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 3.11326607 4.2173286 0.69027452
#>  [2,] 0.57140711 0.2809874 1.23835654
#>  [3,] 1.64044084 6.9506616 1.54414062
#>  [4,] 0.70921879 0.4909384 0.31889740
#>  [5,] 0.79317113 4.0160167 2.46094938
#>  [6,] 1.08714814 0.1885513 0.05889212
#>  [7,] 0.61959217 1.0114227 1.31245608
#>  [8,] 0.25278266 0.4631303 1.43804870
#>  [9,] 0.16353599 2.6443182 1.65427920
#> [10,] 0.05389827 0.1572385 0.50326601

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.108674866 0.6357315 4.3722563
#>  [2,] 0.925576286 1.2572712 1.9231036
#>  [3,] 0.322790029 1.2185307 1.3006475
#>  [4,] 2.600867882 4.7804077 0.2762435
#>  [5,] 0.005031678 0.6234332 0.1848643
#>  [6,] 0.078806465 0.1417904 1.1848863
#>  [7,] 3.327118520 1.2212292 0.2564957
#>  [8,] 0.220754194 1.6862659 2.8158842
#>  [9,] 0.050080558 1.7144888 3.6432400
#> [10,] 0.783528942 2.8071639 0.1571738
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0557196 1.0557196 1.0557196
#>  [2,] 2.8611249 2.8611249 2.8611249
#>  [3,] 0.4525526 0.4525526 0.4525526
#>  [4,] 5.8975663 5.8975663 5.8975663
#>  [5,] 0.7971816 0.7971816 0.7971816
#>  [6,] 0.8070249 0.8070249 0.8070249
#>  [7,] 0.4045939 0.4045939 0.4045939
#>  [8,] 1.6871936 1.6871936 1.6871936
#>  [9,] 0.4656748 0.4656748 0.4656748
#> [10,] 0.1675153 0.1675153 0.1675153
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4023431 0.4980479 2.4148258
#>  [2,] 0.6922990 1.7057521 3.4030735
#>  [3,] 0.1956877 0.6730917 0.7889962
#>  [4,] 2.2213923 1.4634981 1.2433798
#>  [5,] 0.4574220 1.4221485 1.7205650
#>  [6,] 0.7006166 2.5465535 0.7878609
#>  [7,] 0.1200352 0.1676696 5.8321804
#>  [8,] 2.4043148 1.8851310 0.7680328
#>  [9,] 1.7137346 0.2567667 0.2108625
#> [10,] 2.2722601 0.5105574 2.9710462

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 1.01073123 0.5968092 0.32253159
#>  [2,] 1.39417419 0.3916795 1.57798179
#>  [3,] 0.40529806 1.7590759 0.76390298
#>  [4,] 7.21872389 5.0485740 3.41217132
#>  [5,] 0.43289897 1.1360717 0.06958954
#>  [6,] 2.54711914 3.9494539 4.65847243
#>  [7,] 2.91562814 0.6610833 3.98698412
#>  [8,] 0.02337359 2.0572624 0.06084927
#>  [9,] 2.55333812 1.5987681 1.23911967
#> [10,] 0.15374169 1.5366631 0.09826396
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.49517869 0.49517869 0.49517869
#>  [2,] 0.61193747 0.61193747 0.61193747
#>  [3,] 0.93940033 0.93940033 0.93940033
#>  [4,] 0.05710744 0.05710744 0.05710744
#>  [5,] 2.30431360 2.30431360 2.30431360
#>  [6,] 1.58704225 1.58704225 1.58704225
#>  [7,] 0.23062217 0.23062217 0.23062217
#>  [8,] 0.51468518 0.51468518 0.51468518
#>  [9,] 4.84252702 4.84252702 4.84252702
#> [10,] 0.86792106 0.86792106 0.86792106
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>              [,1]      [,2]      [,3]
#>  [1,] 0.634326734 0.5005984 0.6697385
#>  [2,] 2.305046540 1.0921113 1.6373553
#>  [3,] 0.130708128 0.2308279 0.4604235
#>  [4,] 0.039555755 0.7729195 1.2822511
#>  [5,] 1.161567014 1.2242801 1.0098080
#>  [6,] 0.772148962 0.1114476 0.2183045
#>  [7,] 0.913083990 1.8884231 0.7173358
#>  [8,] 0.704603031 3.6099302 0.6287724
#>  [9,] 0.005544441 0.7650614 1.4577350
#> [10,] 0.123145460 0.4637390 0.8867030

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 4.5373814 0.01662521 2.45958414
#>  [2,] 1.0105040 0.12111706 4.52549186
#>  [3,] 0.3773854 0.20879684 0.27905438
#>  [4,] 0.9610277 0.96102775 0.31047129
#>  [5,] 1.1293530 0.41823064 0.23189549
#>  [6,] 0.3975275 0.79064919 2.00691550
#>  [7,] 0.2886121 0.28861209 0.07193835
#>  [8,] 0.8945515 0.09099305 0.59634510
#>  [9,] 0.6170275 1.09960940 1.72354967
#> [10,] 0.1299425 3.05958704 0.89996383
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1778187 0.1778187 0.1778187
#>  [2,] 0.7379442 0.7379442 0.7379442
#>  [3,] 0.6289454 0.6289454 0.6289454
#>  [4,] 1.5499371 1.5499371 1.5499371
#>  [5,] 0.1042668 0.1042668 0.1042668
#>  [6,] 1.0369135 1.0369135 1.0369135
#>  [7,] 1.3638599 1.3638599 1.3638599
#>  [8,] 1.6680577 1.6680577 1.6680577
#>  [9,] 0.2915225 0.2915225 0.2915225
#> [10,] 4.0762215 4.0762215 4.0762215
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 2.77339243 1.37966774 1.58940757
#>  [2,] 0.12440882 0.65994421 0.15326950
#>  [3,] 3.32179360 2.22784582 2.54332437
#>  [4,] 0.05603148 0.34196312 3.76070136
#>  [5,] 1.31854716 0.09445382 0.86988047
#>  [6,] 1.62273749 0.34360819 0.27369524
#>  [7,] 1.03812197 2.67259973 0.09360734
#>  [8,] 0.48417043 0.33608296 0.76301067
#>  [9,] 0.65234788 1.71082148 0.84954761
#> [10,] 0.43019051 1.55483828 0.57671464

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.9313475 3.97045098 1.0304753
#>  [2,] 1.0004395 0.14824986 1.0004395
#>  [3,] 3.9635382 0.78375170 1.7854271
#>  [4,] 2.3016403 2.30164034 2.3016403
#>  [5,] 1.5707792 1.75457404 1.7545740
#>  [6,] 0.7118209 3.57966968 0.0586191
#>  [7,] 0.6739506 0.45027404 1.0902747
#>  [8,] 0.9399408 0.01635684 1.9403990
#>  [9,] 1.2299013 1.22990132 0.6340910
#> [10,] 1.7392624 0.08656703 1.7392624
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 3.67176478 3.67176478 3.67176478
#>  [2,] 0.93186921 0.93186921 0.93186921
#>  [3,] 0.68653295 0.68653295 0.68653295
#>  [4,] 0.53312847 0.53312847 0.53312847
#>  [5,] 1.23507887 1.23507887 1.23507887
#>  [6,] 0.01558058 0.01558058 0.01558058
#>  [7,] 1.62155490 1.62155490 1.62155490
#>  [8,] 1.65696217 1.65696217 1.65696217
#>  [9,] 0.87382043 0.87382043 0.87382043
#> [10,] 0.74751498 0.74751498 0.74751498
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.36424731 2.72865882 0.6282872
#>  [2,] 2.71973469 0.23859569 0.7369221
#>  [3,] 0.43296984 0.38229533 2.1128046
#>  [4,] 0.12871273 0.28861376 5.9233396
#>  [5,] 0.05492062 0.81509629 1.5587229
#>  [6,] 0.94771504 2.79618597 0.5142734
#>  [7,] 0.74332786 0.32066574 2.2317696
#>  [8,] 0.59011194 0.98909483 0.5793500
#>  [9,] 0.96079040 2.62201176 1.8816694
#> [10,] 0.81412558 0.03222875 0.9615392

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.90784393 0.3521905 0.9078439
#>  [2,] 2.31335260 2.8600351 2.8600351
#>  [3,] 0.08011205 1.1999459 2.8130259
#>  [4,] 0.61876415 0.8839213 2.4220790
#>  [5,] 2.87348022 3.0651606 3.0493188
#>  [6,] 5.33895697 3.9440249 1.0786535
#>  [7,] 0.45641791 0.4055321 0.4564179
#>  [8,] 0.68375974 8.0723360 1.3443070
#>  [9,] 0.68060966 1.3929506 0.7258666
#> [10,] 0.49166317 1.4073546 1.3951422
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.14932279 2.14932279 2.14932279
#>  [2,] 1.28665260 1.28665260 1.28665260
#>  [3,] 0.55459261 0.55459261 0.55459261
#>  [4,] 0.05199948 0.05199948 0.05199948
#>  [5,] 1.26283399 1.26283399 1.26283399
#>  [6,] 1.72394223 1.72394223 1.72394223
#>  [7,] 0.73910859 0.73910859 0.73910859
#>  [8,] 0.32421919 0.32421919 0.32421919
#>  [9,] 0.80944559 0.80944559 0.80944559
#> [10,] 0.32278078 0.32278078 0.32278078
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>              [,1]       [,2]      [,3]
#>  [1,] 2.650884313 0.04893264 0.2699623
#>  [2,] 1.137256497 0.39922768 1.1279337
#>  [3,] 0.010786264 0.21428589 0.2443438
#>  [4,] 0.869566069 1.76235258 1.7643096
#>  [5,] 0.669196466 4.87906585 2.6488179
#>  [6,] 1.006040161 0.08478980 0.1335703
#>  [7,] 0.008360733 0.28313404 0.2895630
#>  [8,] 2.246790248 0.67551988 1.8421246
#>  [9,] 0.118440729 2.55522403 0.9883260
#> [10,] 0.948941077 0.48873145 0.3767522

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 2.4104584 2.3374025 0.08721876
#>  [2,] 1.0758109 2.0136905 1.13747882
#>  [3,] 2.4200838 0.2087299 5.49933856
#>  [4,] 2.9639980 5.5489786 0.73578842
#>  [5,] 0.7348333 0.7348333 0.73483334
#>  [6,] 0.3148535 1.3566480 1.35664803
#>  [7,] 1.0087096 3.6522292 1.77896564
#>  [8,] 3.3596949 0.4235807 0.69351613
#>  [9,] 2.7297466 0.8941648 2.72974657
#> [10,] 0.1603909 2.1663664 0.13353797
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>               [,1]         [,2]         [,3]
#>  [1,] 1.2599361109 1.2599361109 1.2599361109
#>  [2,] 0.0003263657 0.0003263657 0.0003263657
#>  [3,] 0.3265229658 0.3265229658 0.3265229658
#>  [4,] 1.0065697617 1.0065697617 1.0065697617
#>  [5,] 0.1386982957 0.1386982957 0.1386982957
#>  [6,] 0.1725123958 0.1725123958 0.1725123958
#>  [7,] 0.1236663382 0.1236663382 0.1236663382
#>  [8,] 4.6605715041 4.6605715041 4.6605715041
#>  [9,] 1.8256693463 1.8256693463 1.8256693463
#> [10,] 0.7092520007 0.7092520007 0.7092520007
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2029320 0.2520558 0.18936949
#>  [2,] 0.6459733 0.1000898 2.42997961
#>  [3,] 0.1709333 0.1410522 0.26784957
#>  [4,] 0.5410022 0.1256482 0.03060632
#>  [5,] 1.7363894 1.0210093 0.84093242
#>  [6,] 0.1402840 0.6174466 0.15754002
#>  [7,] 2.2239973 1.0216543 2.45154967
#>  [8,] 1.4178793 0.3444070 2.00773526
#>  [9,] 2.3796841 0.8628923 0.50562526
#> [10,] 0.1668174 1.4677297 0.10987900

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.86121597 0.8686777 0.26769541
#>  [2,] 0.31117084 2.0667736 0.24612856
#>  [3,] 0.07301227 1.9149960 0.04377015
#>  [4,] 2.90612011 0.1202848 4.09735162
#>  [5,] 0.40420881 2.4793330 3.07057239
#>  [6,] 2.97600468 0.2688616 0.11542768
#>  [7,] 0.02122872 0.2120364 0.21203638
#>  [8,] 0.54751470 3.0107742 3.01077423
#>  [9,] 2.80885652 0.3591569 0.54389840
#> [10,] 0.52062275 0.4423166 2.57606976
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9723847 1.9723847 1.9723847
#>  [2,] 1.7152129 1.7152129 1.7152129
#>  [3,] 4.5949824 4.5949824 4.5949824
#>  [4,] 0.7651686 0.7651686 0.7651686
#>  [5,] 0.1867918 0.1867918 0.1867918
#>  [6,] 0.9955636 0.9955636 0.9955636
#>  [7,] 1.5371379 1.5371379 1.5371379
#>  [8,] 3.7289702 3.7289702 3.7289702
#>  [9,] 0.8523991 0.8523991 0.8523991
#> [10,] 0.1578420 0.1578420 0.1578420
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]      [,2]        [,3]
#>  [1,] 0.5180743 0.2962175 0.083674957
#>  [2,] 2.4132543 0.1182137 0.004840716
#>  [3,] 0.5873550 1.8643842 1.017496981
#>  [4,] 1.3428810 1.1780410 0.279103749
#>  [5,] 2.0901723 1.5236476 0.708594809
#>  [6,] 1.2016425 1.0372933 0.012719229
#>  [7,] 0.2660846 1.0367305 0.730274803
#>  [8,] 0.3868298 2.0838429 0.151765356
#>  [9,] 0.4019550 1.9712896 0.624029197
#> [10,] 0.7694234 1.0416961 2.853865522

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.26781753 0.02219473 0.17938190
#>  [2,] 1.27402004 1.41510769 0.53248776
#>  [3,] 0.02198795 4.04967624 5.35203955
#>  [4,] 0.04320402 0.04320402 0.04320402
#>  [5,] 1.55744696 0.79222700 0.33339922
#>  [6,] 4.04362581 1.64886481 2.29229194
#>  [7,] 0.17416638 0.19318363 0.19318363
#>  [8,] 1.27899418 0.72798675 0.42641615
#>  [9,] 1.88949130 1.88949130 0.60032358
#> [10,] 6.62676563 0.72529319 0.66028843
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7569855 2.7569855 2.7569855
#>  [2,] 0.6076764 0.6076764 0.6076764
#>  [3,] 0.2952659 0.2952659 0.2952659
#>  [4,] 1.1710997 1.1710997 1.1710997
#>  [5,] 0.3536897 0.3536897 0.3536897
#>  [6,] 3.4098422 3.4098422 3.4098422
#>  [7,] 0.1623595 0.1623595 0.1623595
#>  [8,] 0.3996127 0.3996127 0.3996127
#>  [9,] 0.5507121 0.5507121 0.5507121
#> [10,] 0.7174058 0.7174058 0.7174058
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.72092780 0.6353683 0.1885697
#>  [2,] 1.12707634 0.4857830 3.0250350
#>  [3,] 0.06980892 2.0410058 0.2802057
#>  [4,] 0.01580753 0.4348212 2.5094939
#>  [5,] 1.02528308 0.1931102 0.2318550
#>  [6,] 1.42517705 0.2658737 1.7916301
#>  [7,] 1.03920118 2.5048524 0.4739816
#>  [8,] 0.13473537 1.0640213 0.1857443
#>  [9,] 0.07209069 1.8716511 0.6262682
#> [10,] 0.05669899 0.1195259 0.1189821

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.1241378 1.12413783 1.12413783
#>  [2,] 1.8687074 1.61387372 1.86870736
#>  [3,] 1.6402577 1.70220889 0.06719214
#>  [4,] 0.1704999 0.24206103 0.24206103
#>  [5,] 1.0496606 1.07165662 1.07165662
#>  [6,] 0.2940409 1.27729036 0.42648472
#>  [7,] 0.7623963 0.28636970 0.11284331
#>  [8,] 0.4571556 0.47640224 0.42844853
#>  [9,] 0.5142345 0.17111551 1.52000445
#> [10,] 0.6636962 0.09994488 0.09994488
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>               [,1]       [,2]       [,3]
#>  [1,] 0.0604961495 0.09318807 0.06049615
#>  [2,] 0.0002800532 0.24990990 0.01126320
#>  [3,] 0.7951784284 0.38995071 0.38995071
#>  [4,] 0.1225821400 0.07462371 1.87780767
#>  [5,] 0.2005302773 0.79568016 0.17685636
#>  [6,] 0.3725811600 0.09128205 0.09128205
#>  [7,] 0.5726344016 0.18958650 1.45717691
#>  [8,] 1.8876141470 0.41001591 0.22153429
#>  [9,] 0.2410843511 0.24108435 0.24108435
#> [10,] 0.4975219112 2.23375659 1.30584524
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.92518822 0.36518102 0.13129723
#>  [2,] 1.06421739 0.16391176 2.50518753
#>  [3,] 1.26496867 0.09285656 0.79337494
#>  [4,] 0.09172137 1.64125255 0.09172137
#>  [5,] 0.47369585 0.59657297 0.24882541
#>  [6,] 0.16806539 0.08729651 1.99409565
#>  [7,] 0.80734345 0.65905409 0.07455726
#>  [8,] 0.39384294 0.39384294 0.51797880
#>  [9,] 2.01307235 0.36614138 1.79437172
#> [10,] 0.07415509 0.29355892 0.07415509
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.34080415 0.34080415 0.3408041
#>  [2,] 1.03494052 1.67134020 0.1782466
#>  [3,] 1.82631889 0.08590398 1.8263189
#>  [4,] 0.52193587 0.25399064 0.2713307
#>  [5,] 0.45699149 0.98685265 1.0990569
#>  [6,] 0.05175586 0.20452842 0.8103484
#>  [7,] 2.94786189 1.70203714 2.2646565
#>  [8,] 0.73471890 0.01204002 0.6449381
#>  [9,] 0.87231738 0.87231738 0.2433479
#> [10,] 0.43089875 0.43089875 0.4308987
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.5385653 0.8549764 1.23603591
#>  [2,] 0.3147869 1.2018955 1.94435256
#>  [3,] 0.2053280 0.2053280 0.20532800
#>  [4,] 0.3451988 0.1764142 0.17641422
#>  [5,] 0.2653086 0.2653086 0.24803981
#>  [6,] 1.1912399 0.1953015 0.19530152
#>  [7,] 0.9138753 0.4822880 0.91387527
#>  [8,] 0.9116275 0.8491485 0.87475205
#>  [9,] 0.5473238 1.8196458 0.54732381
#> [10,] 0.3577141 0.9349061 0.04644645

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.9068150 0.16074749 3.38452052
#>  [2,] 0.8640693 0.29832648 0.59335548
#>  [3,] 2.1928254 1.28511248 5.69123739
#>  [4,] 0.4948328 0.69354772 0.09566399
#>  [5,] 0.4678031 0.30802774 4.05958312
#>  [6,] 1.0578784 0.01434318 0.55565416
#>  [7,] 2.2422518 0.03766847 0.82607731
#>  [8,] 4.2543074 0.08820594 3.30382912
#>  [9,] 0.2728900 1.78045825 2.07433186
#> [10,] 1.6319414 1.76862348 1.76862348
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.72818658 1.01265063 1.38020935
#>  [2,] 3.75522593 3.75522593 3.05365939
#>  [3,] 0.18637832 1.91181848 0.14623798
#>  [4,] 2.43006751 2.10741315 1.59149674
#>  [5,] 0.76068691 4.39504712 1.94557200
#>  [6,] 0.08206797 0.06838769 0.02860557
#>  [7,] 1.16549590 1.52675466 0.63076016
#>  [8,] 4.78572297 0.52870122 0.81801360
#>  [9,] 1.57954379 1.06401186 2.24444573
#> [10,] 0.43194148 3.97976466 2.99293406
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8646249 0.2725060 1.5419934
#>  [2,] 2.3232381 1.4605166 1.5528455
#>  [3,] 1.2447897 2.7510546 4.6973400
#>  [4,] 2.9158747 0.1607615 0.3974157
#>  [5,] 0.9923565 0.0629623 0.9923565
#>  [6,] 3.3587283 1.5654712 2.2088207
#>  [7,] 4.2590463 0.8893089 0.5072222
#>  [8,] 1.4139694 0.8264337 1.4139694
#>  [9,] 0.6718665 0.3182049 0.8453354
#> [10,] 1.0282729 1.0282729 1.0282729
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.57630930 0.88404494 2.75857107
#>  [2,] 0.17365417 0.17365417 0.17365417
#>  [3,] 0.80244144 0.80244144 0.80244144
#>  [4,] 1.55128812 0.06546716 0.91502085
#>  [5,] 0.76386816 1.23358711 4.34933423
#>  [6,] 0.88019420 0.16709649 0.48326430
#>  [7,] 0.09974069 0.09974069 0.09974069
#>  [8,] 1.86566917 1.74516823 1.86566917
#>  [9,] 3.91233986 3.91233986 3.43034374
#> [10,] 1.27343662 0.13588753 1.19903485
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.09729492 3.7504525 2.2532447
#>  [2,] 1.75005674 0.4009548 1.7500567
#>  [3,] 0.67341195 2.9207404 5.0306818
#>  [4,] 0.46890749 0.4689075 0.4689075
#>  [5,] 3.53992901 1.1527170 1.7765354
#>  [6,] 1.87744290 0.9731358 0.7622396
#>  [7,] 3.30716568 3.3071657 2.3017147
#>  [8,] 0.48480941 0.1286952 3.1063844
#>  [9,] 0.06610248 0.2614477 2.3758828
#> [10,] 0.53083566 0.7448904 1.2131203

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.41230765 0.56219304 1.31616274
#>  [2,] 0.14185624 1.54195588 1.81960837
#>  [3,] 0.18008455 0.18008455 0.61172033
#>  [4,] 0.56233706 0.04377296 0.53625671
#>  [5,] 0.35851722 0.65256053 0.65256053
#>  [6,] 0.17138580 0.17138580 0.17138580
#>  [7,] 0.08119726 0.08119726 0.08119726
#>  [8,] 0.01434861 0.82687488 2.08478243
#>  [9,] 0.99109634 0.08552085 0.38047148
#> [10,] 0.29764439 0.29764439 0.64348890
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.00768202 0.00768202 0.00768202
#>  [2,] 0.02139301 0.13277326 0.13277326
#>  [3,] 0.21033000 0.21033000 0.28219648
#>  [4,] 0.17616915 0.17616915 0.17616915
#>  [5,] 0.20267437 0.20267437 0.20267437
#>  [6,] 0.04720855 0.37147011 0.04720855
#>  [7,] 0.23850568 0.20458387 0.20458387
#>  [8,] 0.16971807 0.16971807 0.16971807
#>  [9,] 0.64853908 0.18498433 1.37288424
#> [10,] 0.29101145 0.29101145 0.29101145
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>            [,1]        [,2]      [,3]
#>  [1,] 0.6479793 0.647979341 0.6479793
#>  [2,] 0.8607878 0.860787781 0.8607878
#>  [3,] 1.3801278 0.173325433 0.1286969
#>  [4,] 1.7068852 1.706885185 1.6533252
#>  [5,] 0.1799485 0.179948467 0.1799485
#>  [6,] 0.1146489 0.114648882 0.1146489
#>  [7,] 0.2744805 0.274480535 0.2562429
#>  [8,] 0.1109621 0.005741363 0.7290970
#>  [9,] 0.2117916 0.211791578 0.2117916
#> [10,] 0.1841812 0.184181231 0.8559197
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.27444041 1.64247603 0.41778758
#>  [2,] 0.09720306 0.61786559 0.61786559
#>  [3,] 0.45554342 0.81294217 1.47115831
#>  [4,] 0.70702205 0.70702205 0.70702205
#>  [5,] 0.66830355 0.66984965 0.04603805
#>  [6,] 0.39168904 0.45150554 0.45150554
#>  [7,] 0.14849452 0.14849452 0.14849452
#>  [8,] 0.11552540 0.11552540 0.11552540
#>  [9,] 0.39738283 0.39738283 0.15501177
#> [10,] 0.02378009 0.02378009 0.02378009
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20618308 0.20618308 0.20618308
#>  [2,] 0.11617646 0.05850825 0.04273586
#>  [3,] 0.12581214 0.12581214 0.12581214
#>  [4,] 1.05733665 1.05733665 1.05733665
#>  [5,] 0.04272251 0.22998942 0.22998942
#>  [6,] 0.76615263 0.24503577 1.36669551
#>  [7,] 1.34559882 0.07856804 0.86439895
#>  [8,] 0.36857355 0.36857355 0.43617267
#>  [9,] 1.57499446 1.61327223 1.61327223
#> [10,] 0.85860077 0.85860077 0.85860077

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.46710241 0.31598081 0.11801307
#>  [2,] 0.26872911 0.53045819 0.79534497
#>  [3,] 0.25437460 0.09975080 0.25437460
#>  [4,] 0.40850832 0.65302600 0.40850832
#>  [5,] 0.11989817 0.18656957 0.18656957
#>  [6,] 0.05257804 0.05833075 0.05257804
#>  [7,] 0.16614622 0.16614622 0.16614622
#>  [8,] 0.20254687 1.06012550 0.20254687
#>  [9,] 1.04704406 0.50835301 1.19373699
#> [10,] 0.29424328 0.29424328 0.29424328
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.68632784 0.68632784 0.48843159
#>  [2,] 0.82081111 0.58991836 0.62698149
#>  [3,] 0.04515224 0.96273108 0.96273108
#>  [4,] 0.16967365 0.04980957 0.02688370
#>  [5,] 0.15036022 0.18486653 0.18486653
#>  [6,] 0.17046122 0.17046122 0.17046122
#>  [7,] 0.12171851 0.79789379 0.02367122
#>  [8,] 0.29042461 0.16105626 0.16105626
#>  [9,] 0.26484562 0.23487219 0.23487219
#> [10,] 0.30503491 1.00353735 1.00353735
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.19806546 0.69990637 0.9378884
#>  [2,] 0.07518223 0.08504723 0.2012973
#>  [3,] 0.40976078 1.55394732 0.1294560
#>  [4,] 0.16051972 0.16051972 0.1605197
#>  [5,] 0.30032859 0.30032859 0.3003286
#>  [6,] 0.30223264 0.02569097 0.3898355
#>  [7,] 0.59920189 0.33577214 0.1225502
#>  [8,] 0.24341612 0.19410530 0.2453377
#>  [9,] 0.02323559 0.17035680 0.2015122
#> [10,] 0.08230308 0.44458614 0.2698350
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42407783 0.19464567 1.15666377
#>  [2,] 0.32456851 0.34453995 0.34453995
#>  [3,] 1.25720827 0.29510688 0.29510688
#>  [4,] 0.08458444 1.57229186 0.08503308
#>  [5,] 0.21370770 0.21370770 0.21370770
#>  [6,] 0.81217309 0.88790704 0.68145256
#>  [7,] 0.07936897 0.07936897 0.07936897
#>  [8,] 0.21741636 0.79776279 0.72334697
#>  [9,] 0.02932196 0.02932196 0.02932196
#> [10,] 0.08406722 0.08406722 0.08406722

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05352596 0.49016942 0.51324771
#>  [2,] 0.12779880 0.06112272 0.06112272
#>  [3,] 0.22807955 1.39680926 0.72604913
#>  [4,] 0.49904806 0.19573657 0.02082716
#>  [5,] 0.37294233 0.37294233 0.40836480
#>  [6,] 0.41345886 0.29073960 0.29073960
#>  [7,] 0.24176605 0.22763138 0.28347682
#>  [8,] 0.24113262 0.24113262 0.24113262
#>  [9,] 0.26208532 0.74400878 0.26208532
#> [10,] 0.58552621 0.44100683 0.85216836
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.12025541 0.12025541 0.44657019
#>  [2,] 0.29705972 0.26235920 0.26235920
#>  [3,] 0.36696871 0.62576558 0.62576558
#>  [4,] 0.42568800 0.72302445 0.01169305
#>  [5,] 0.20092088 0.20092088 0.04567158
#>  [6,] 1.88817616 0.12324489 0.12932269
#>  [7,] 0.22921481 0.22921481 0.22921481
#>  [8,] 0.07529077 0.22550470 0.22042588
#>  [9,] 0.07593922 0.07593922 0.07593922
#> [10,] 0.19037078 0.19037078 0.19037078
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.13125477 0.699073694 0.080885816
#>  [2,] 1.24060190 0.951152815 0.195296699
#>  [3,] 0.06450293 0.064502929 0.064502929
#>  [4,] 0.05825351 0.366510105 0.007522907
#>  [5,] 0.15926339 0.159802220 0.199243372
#>  [6,] 0.05382444 1.115774674 0.053824439
#>  [7,] 0.44573163 0.087732905 1.743378877
#>  [8,] 0.54156252 0.896495545 0.169241207
#>  [9,] 1.79354130 0.684923670 0.677843842
#> [10,] 0.26641578 0.008864125 0.424025304
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.902989171 0.060399199 0.846964173
#>  [2,] 0.002628612 0.002628612 0.002628612
#>  [3,] 0.263102974 0.540001651 0.540001651
#>  [4,] 0.490804790 0.422390436 0.422390436
#>  [5,] 0.257902999 0.737374340 0.222063710
#>  [6,] 0.106074369 0.650371560 0.650371560
#>  [7,] 0.077878721 0.515207034 0.239865046
#>  [8,] 0.709371044 0.063379959 0.973634977
#>  [9,] 0.186747555 0.186747555 1.064214891
#> [10,] 0.249937066 0.363134182 0.221142995

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.45130463 0.62051810 0.21432610
#>  [2,] 0.68986142 0.51122028 0.09714103
#>  [3,] 0.14296996 0.03669042 0.18338292
#>  [4,] 0.10394366 0.10394366 0.02542832
#>  [5,] 0.14573746 0.03628600 0.12955143
#>  [6,] 0.14653615 0.14653615 0.14653615
#>  [7,] 0.09042104 0.42072705 0.42072705
#>  [8,] 0.11659464 0.11659464 0.01445020
#>  [9,] 0.03992886 0.03992886 0.35077891
#> [10,] 0.07016976 0.13213085 0.13213085
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2159689 0.8316512 0.76546544
#>  [2,] 0.5920432 0.5920432 0.59204318
#>  [3,] 0.1118481 0.5035394 0.08948169
#>  [4,] 0.1930228 0.3656930 0.53222290
#>  [5,] 1.2361565 1.3204534 1.32045343
#>  [6,] 0.8745992 0.2542259 0.05595578
#>  [7,] 0.7906592 0.3677211 0.17831520
#>  [8,] 0.2188953 0.2188953 0.21889531
#>  [9,] 0.2109445 0.2109445 0.21094454
#> [10,] 0.8086147 0.6065675 0.78633813
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.02792477 0.529731613 0.276203515
#>  [2,] 0.11139490 0.111394902 0.111394902
#>  [3,] 0.26655062 0.266550618 0.266550618
#>  [4,] 0.26537442 0.265374419 0.265374419
#>  [5,] 0.36921508 0.626647407 0.626647407
#>  [6,] 0.16995812 0.187621812 0.187621812
#>  [7,] 0.13119569 1.552521339 0.403045859
#>  [8,] 0.77326546 0.009297909 0.009297909
#>  [9,] 0.02631830 0.483380056 0.483380056
#> [10,] 0.33391599 0.471712597 0.471712597
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3601622 0.05989741 0.36016216
#>  [2,] 0.1569390 0.40103637 0.12912424
#>  [3,] 0.4890967 0.04901784 0.04901784
#>  [4,] 0.3550749 0.70639644 0.35507489
#>  [5,] 0.4358715 0.03844146 0.11240739
#>  [6,] 0.2340987 0.23409870 0.23409870
#>  [7,] 0.2751530 0.27862437 0.35778433
#>  [8,] 0.6964047 1.24633195 1.46386462
#>  [9,] 1.2785406 0.88656160 0.77366334
#> [10,] 0.4328939 0.43289391 0.06620005