Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3059180 1.6300461 0.8624596
#>  [2,] 1.6135009 1.6135009 0.5563823
#>  [3,] 0.4639118 5.5722456 1.9062320
#>  [4,] 2.9105793 0.8647398 2.8016570
#>  [5,] 0.6854204 0.6854204 0.6854204
#>  [6,] 0.5478285 7.1967752 0.7323045
#>  [7,] 1.4366422 2.5573822 0.5633527
#>  [8,] 3.2917286 0.2124931 4.1231670
#>  [9,] 1.7092488 1.4128799 3.5137240
#> [10,] 0.3516861 1.6112115 0.7229600
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4148717 0.4148717 0.4148717
#>  [2,] 5.9362110 5.9362110 5.9362110
#>  [3,] 0.7913332 0.7913332 0.7913332
#>  [4,] 0.9503529 0.9503529 0.9503529
#>  [5,] 1.2724723 1.2724723 1.2724723
#>  [6,] 2.6725916 2.6725916 2.6725916
#>  [7,] 0.5555901 0.5555901 0.5555901
#>  [8,] 3.1010583 3.1010583 3.1010583
#>  [9,] 0.5898840 0.5898840 0.5898840
#> [10,] 1.6046976 1.6046976 1.6046976
## independence
rpextmo(10, 3, b = 1)
#>            [,1]        [,2]      [,3]
#>  [1,] 0.2365313 0.705273081 1.6589202
#>  [2,] 0.7774296 0.573370717 0.1610316
#>  [3,] 1.1712198 4.932321535 0.5330414
#>  [4,] 0.0787795 3.939988976 0.4498867
#>  [5,] 0.1830043 1.659614797 0.2854120
#>  [6,] 0.1155532 0.001704002 2.4874488
#>  [7,] 0.7419089 0.488818648 0.7451319
#>  [8,] 0.6583169 0.617412459 0.6369775
#>  [9,] 0.1647628 0.786742627 1.2247204
#> [10,] 0.2911229 0.467807916 0.1400988

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 3.42851214 4.06478997 1.2935722
#>  [2,] 3.00473240 1.78627835 2.1999690
#>  [3,] 0.52526495 2.48391157 1.6205068
#>  [4,] 0.75441461 2.82637376 5.1259463
#>  [5,] 1.09756256 0.29026773 2.5654596
#>  [6,] 1.23647943 1.36847258 0.2882433
#>  [7,] 0.22433234 0.22433234 0.2243323
#>  [8,] 0.25327621 1.02597055 1.6879344
#>  [9,] 2.30529826 0.96503966 2.2884972
#> [10,] 0.05062043 0.02427085 2.3042018
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.80212590 0.80212590 0.80212590
#>  [2,] 0.81074282 0.81074282 0.81074282
#>  [3,] 1.45410878 1.45410878 1.45410878
#>  [4,] 0.69315677 0.69315677 0.69315677
#>  [5,] 0.48808954 0.48808954 0.48808954
#>  [6,] 0.91767580 0.91767580 0.91767580
#>  [7,] 0.48842997 0.48842997 0.48842997
#>  [8,] 0.32566918 0.32566918 0.32566918
#>  [9,] 1.04256244 1.04256244 1.04256244
#> [10,] 0.02020792 0.02020792 0.02020792
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.48836961 0.90369759 0.8184415
#>  [2,] 0.81466978 0.35112175 2.3132029
#>  [3,] 0.18598371 0.07961135 0.7304132
#>  [4,] 1.43854298 1.42651148 2.4765903
#>  [5,] 4.38104814 0.91615815 1.6636460
#>  [6,] 0.15964929 0.49530190 0.1715667
#>  [7,] 1.08575640 0.86460388 1.5289857
#>  [8,] 0.68252533 1.31954031 0.2443551
#>  [9,] 0.09856769 0.92252984 0.4097638
#> [10,] 2.02091467 2.05565375 2.3350439

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.425562304 0.42556230 0.42556230
#>  [2,] 0.080170594 0.08017059 0.08017059
#>  [3,] 1.199747729 2.82015072 3.12418016
#>  [4,] 1.335754400 2.63203124 2.63203124
#>  [5,] 0.005575608 1.61700989 0.42034383
#>  [6,] 0.338480829 2.87108967 2.41427338
#>  [7,] 4.708523360 2.54463406 1.42040359
#>  [8,] 0.498996255 1.61265714 1.10950020
#>  [9,] 0.939716708 0.61586737 0.93971671
#> [10,] 0.747315422 1.80490218 1.37560974
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2155630 0.2155630 0.2155630
#>  [2,] 0.1735562 0.1735562 0.1735562
#>  [3,] 0.4700880 0.4700880 0.4700880
#>  [4,] 0.9415514 0.9415514 0.9415514
#>  [5,] 0.3635807 0.3635807 0.3635807
#>  [6,] 1.6441541 1.6441541 1.6441541
#>  [7,] 0.7591997 0.7591997 0.7591997
#>  [8,] 0.9019686 0.9019686 0.9019686
#>  [9,] 3.0722247 3.0722247 3.0722247
#> [10,] 0.8342096 0.8342096 0.8342096
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.03161520 0.6209930 0.1005191
#>  [2,] 0.89742709 0.7055037 0.7342058
#>  [3,] 0.71728630 0.7079265 1.5563447
#>  [4,] 0.11642327 0.8009777 0.4814070
#>  [5,] 1.87237151 0.4128227 0.6275572
#>  [6,] 0.28591178 0.8271020 0.2360194
#>  [7,] 0.60142451 0.7732999 0.9318872
#>  [8,] 0.78272872 1.0779593 1.2474261
#>  [9,] 0.07653281 4.7834716 0.9346616
#> [10,] 1.37806429 0.7712028 0.3303511

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0248657 2.0248657 0.2383915
#>  [2,] 1.0608876 0.1268337 0.3372787
#>  [3,] 1.2534631 1.0924807 1.1234089
#>  [4,] 1.1277298 0.6239482 3.7576449
#>  [5,] 0.5987208 0.5987208 0.5987208
#>  [6,] 3.2588101 2.3333670 0.1229977
#>  [7,] 0.1011620 1.1817376 0.2914624
#>  [8,] 0.7737221 1.2328454 1.2328454
#>  [9,] 3.0293237 2.4194823 7.0734119
#> [10,] 5.7720465 0.8678799 0.5566739
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.1869110 2.1869110 2.1869110
#>  [2,] 0.2526761 0.2526761 0.2526761
#>  [3,] 0.5484514 0.5484514 0.5484514
#>  [4,] 0.7994927 0.7994927 0.7994927
#>  [5,] 2.4446709 2.4446709 2.4446709
#>  [6,] 0.1052044 0.1052044 0.1052044
#>  [7,] 0.1519141 0.1519141 0.1519141
#>  [8,] 0.3990969 0.3990969 0.3990969
#>  [9,] 0.5518345 0.5518345 0.5518345
#> [10,] 0.9175344 0.9175344 0.9175344
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.3625069 0.4020118 0.18235072
#>  [2,] 0.9535181 0.4793429 0.94442966
#>  [3,] 0.2337625 0.4875483 3.49381311
#>  [4,] 0.3231089 0.7053599 0.22302844
#>  [5,] 1.1831813 1.9980133 0.05390188
#>  [6,] 0.8879945 0.1198635 0.36315108
#>  [7,] 0.8997538 1.0027558 0.13604007
#>  [8,] 0.5812409 1.7617766 0.87664608
#>  [9,] 0.3452149 4.0385555 2.15280486
#> [10,] 1.0832552 2.0761850 1.01265634

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.7980572 0.7980572 0.79805725
#>  [2,] 3.2230504 2.8220129 2.92086744
#>  [3,] 0.7625991 0.2051375 1.98574404
#>  [4,] 2.2359992 1.3543594 0.50533602
#>  [5,] 0.3892673 0.7553390 0.75533898
#>  [6,] 0.2588140 0.7038372 0.99855752
#>  [7,] 4.0040812 4.9673796 2.06629994
#>  [8,] 0.2953188 3.4586164 5.39532950
#>  [9,] 1.3762603 0.1010491 0.06045423
#> [10,] 0.4663113 1.7909390 1.36030814
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.20728055 1.20728055 1.20728055
#>  [2,] 0.03243259 0.03243259 0.03243259
#>  [3,] 0.60644585 0.60644585 0.60644585
#>  [4,] 5.77615514 5.77615514 5.77615514
#>  [5,] 0.35782140 0.35782140 0.35782140
#>  [6,] 0.25221837 0.25221837 0.25221837
#>  [7,] 1.38513429 1.38513429 1.38513429
#>  [8,] 0.30360440 0.30360440 0.30360440
#>  [9,] 3.20983867 3.20983867 3.20983867
#> [10,] 1.76578619 1.76578619 1.76578619
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2512989 0.6422247 1.48618064
#>  [2,] 0.4381636 0.7648067 0.80573084
#>  [3,] 1.8158595 0.4868860 1.54680111
#>  [4,] 0.5383960 1.4398605 2.28846732
#>  [5,] 0.4738206 0.4352889 4.16657152
#>  [6,] 0.9353373 0.2482704 0.68563039
#>  [7,] 0.7260697 1.2197521 0.90689958
#>  [8,] 0.3065514 1.3890264 0.08576331
#>  [9,] 1.1174384 3.1633774 0.97636567
#> [10,] 0.9452451 1.2088034 0.23679241

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>               [,1]      [,2]      [,3]
#>  [1,] 1.3066804269 1.3066804 1.3066804
#>  [2,] 3.5723987126 0.2551847 0.8303419
#>  [3,] 0.5283302595 2.2146153 0.2422626
#>  [4,] 2.0848157217 1.3131743 1.9467660
#>  [5,] 0.0005935809 0.7086861 1.8709257
#>  [6,] 7.4937622338 2.7497778 0.6637653
#>  [7,] 0.5533475296 1.1194697 0.2209571
#>  [8,] 1.2239684386 1.2239684 0.3550081
#>  [9,] 3.1249422859 3.1249423 3.0453662
#> [10,] 1.6844190023 0.6360915 2.7223290
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.39185153 0.39185153 0.39185153
#>  [2,] 0.09437172 0.09437172 0.09437172
#>  [3,] 1.24551622 1.24551622 1.24551622
#>  [4,] 0.73827333 0.73827333 0.73827333
#>  [5,] 0.59665810 0.59665810 0.59665810
#>  [6,] 0.05525057 0.05525057 0.05525057
#>  [7,] 1.18563510 1.18563510 1.18563510
#>  [8,] 0.96656796 0.96656796 0.96656796
#>  [9,] 0.30571102 0.30571102 0.30571102
#> [10,] 2.83325472 2.83325472 2.83325472
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.03813705 0.6063700 0.5230508
#>  [2,] 0.56400767 0.7693581 0.5652916
#>  [3,] 0.82572652 1.6725279 0.1721555
#>  [4,] 0.16839532 1.0784148 0.3446646
#>  [5,] 1.27307956 0.0936121 0.2787621
#>  [6,] 1.51577393 0.8792411 0.4193592
#>  [7,] 0.97180721 1.7806465 1.6869065
#>  [8,] 0.06821595 0.5039852 1.2641816
#>  [9,] 1.62416481 1.2073019 1.1169346
#> [10,] 1.37249796 1.9573499 0.3944405

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 2.2069568 2.6306707 2.63067070
#>  [2,] 2.9291986 1.4746028 0.91582615
#>  [3,] 6.5971526 4.7133869 1.10981877
#>  [4,] 0.1371317 0.1226198 1.56446847
#>  [5,] 4.5840626 0.5994617 2.65033747
#>  [6,] 0.2541133 1.0064080 1.00640798
#>  [7,] 2.0854339 1.0212878 4.15852100
#>  [8,] 1.8313363 1.8313363 1.37772223
#>  [9,] 1.0886834 1.1337645 0.05490003
#> [10,] 1.4510659 4.4450978 0.89002664
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23924073 0.23924073 0.23924073
#>  [2,] 0.38511990 0.38511990 0.38511990
#>  [3,] 1.73151503 1.73151503 1.73151503
#>  [4,] 0.08044279 0.08044279 0.08044279
#>  [5,] 1.35973543 1.35973543 1.35973543
#>  [6,] 3.17173377 3.17173377 3.17173377
#>  [7,] 3.39092339 3.39092339 3.39092339
#>  [8,] 0.51230966 0.51230966 0.51230966
#>  [9,] 0.30960961 0.30960961 0.30960961
#> [10,] 1.22646357 1.22646357 1.22646357
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1736812 0.49444607 1.2867584
#>  [2,] 1.0844226 1.90363870 0.2193549
#>  [3,] 0.1804287 0.07453863 1.3217524
#>  [4,] 0.2101953 3.04735275 0.3078331
#>  [5,] 1.0916156 0.32734853 0.1294015
#>  [6,] 1.5615724 7.50210855 4.1361803
#>  [7,] 1.4864429 3.60450712 0.3244015
#>  [8,] 0.6592006 1.83048853 1.1186424
#>  [9,] 0.1355371 2.92461446 0.4482271
#> [10,] 0.9493083 0.56884334 0.8172412

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3850984 1.9163715 1.9163715
#>  [2,] 0.4885468 0.4885468 0.4885468
#>  [3,] 0.3155116 2.8984527 3.8949667
#>  [4,] 0.4298024 0.8789917 0.9656007
#>  [5,] 1.4254410 2.2315519 2.4486693
#>  [6,] 1.1025046 1.1025212 1.3692467
#>  [7,] 1.4987756 1.8577946 0.4648777
#>  [8,] 1.0938149 1.0938149 0.4089897
#>  [9,] 1.3774431 2.1081764 1.2547818
#> [10,] 0.9309627 0.9309627 0.9309627
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.83390235 0.83390235 0.83390235
#>  [2,] 3.23809032 3.23809032 3.23809032
#>  [3,] 0.09516422 0.09516422 0.09516422
#>  [4,] 0.99141835 0.99141835 0.99141835
#>  [5,] 0.05179412 0.05179412 0.05179412
#>  [6,] 0.39694556 0.39694556 0.39694556
#>  [7,] 0.32699703 0.32699703 0.32699703
#>  [8,] 1.54924392 1.54924392 1.54924392
#>  [9,] 1.75894702 1.75894702 1.75894702
#> [10,] 0.38229178 0.38229178 0.38229178
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.18231158 2.37957587 0.61581739
#>  [2,] 2.47892535 0.06032215 1.30237565
#>  [3,] 1.29382426 1.31386997 0.31303546
#>  [4,] 0.70640075 1.80625184 1.48892198
#>  [5,] 0.06921099 0.32116033 0.91643636
#>  [6,] 0.75441813 1.62696002 0.40127208
#>  [7,] 0.19130876 0.14334153 2.90629915
#>  [8,] 0.28129473 1.35668484 0.97651401
#>  [9,] 2.79511514 0.57671089 0.66513731
#> [10,] 0.03318889 0.29632622 0.06572415

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.11576445 3.6464774 1.5964543
#>  [2,] 1.36215139 0.4290897 0.4503257
#>  [3,] 0.93999825 0.4038169 0.5632153
#>  [4,] 0.34468615 0.3446862 0.3446862
#>  [5,] 0.82409402 0.3355364 1.1255719
#>  [6,] 1.91916456 1.4255312 0.5665792
#>  [7,] 0.57280040 0.5728004 0.5728004
#>  [8,] 0.51500435 1.3239655 0.9585810
#>  [9,] 0.05303545 1.7451301 1.0568345
#> [10,] 0.26384024 0.7615151 0.7615151
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.09377170 1.09377170 1.09377170
#>  [2,] 0.23313520 0.23313520 0.23313520
#>  [3,] 0.09252142 0.09252142 0.09252142
#>  [4,] 0.08046288 0.08046288 0.08046288
#>  [5,] 4.31901599 4.31901599 4.31901599
#>  [6,] 3.30017897 3.30017897 3.30017897
#>  [7,] 0.40924812 0.40924812 0.40924812
#>  [8,] 1.21001116 1.21001116 1.21001116
#>  [9,] 0.07330407 0.07330407 0.07330407
#> [10,] 0.54994252 0.54994252 0.54994252
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01608883 0.76836196 3.20798552
#>  [2,] 2.54225985 0.55736181 3.21635858
#>  [3,] 0.63078656 0.14300932 0.41661412
#>  [4,] 0.01916496 0.02113391 0.69508645
#>  [5,] 0.08161354 0.66278656 0.00209658
#>  [6,] 2.97977315 1.36523716 0.77470897
#>  [7,] 0.55206659 0.36205300 0.14414761
#>  [8,] 1.14745230 1.74048846 0.24021541
#>  [9,] 1.74761915 1.25641051 0.91536246
#> [10,] 3.30753393 1.35726001 1.07416226

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.83582643 0.99675553 0.24185600
#>  [2,] 0.26030054 0.58199716 0.58199716
#>  [3,] 1.99386690 0.35830381 1.40914944
#>  [4,] 0.07062435 0.07062435 0.07062435
#>  [5,] 1.37547449 1.37547449 0.80034294
#>  [6,] 1.62902528 1.02271563 6.48009217
#>  [7,] 0.97478004 4.63385881 0.13807133
#>  [8,] 1.06760889 0.79737550 1.65233266
#>  [9,] 6.53584536 0.98236958 1.60421247
#> [10,] 0.98044331 0.51445021 0.28156672
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.973482845 1.973482845 1.973482845
#>  [2,] 1.074497669 1.074497669 1.074497669
#>  [3,] 1.209071362 1.209071362 1.209071362
#>  [4,] 0.663309633 0.663309633 0.663309633
#>  [5,] 0.525561317 0.525561317 0.525561317
#>  [6,] 0.228059701 0.228059701 0.228059701
#>  [7,] 0.001606452 0.001606452 0.001606452
#>  [8,] 1.487315694 1.487315694 1.487315694
#>  [9,] 0.003385293 0.003385293 0.003385293
#> [10,] 1.799749790 1.799749790 1.799749790
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4173674 0.26176464 2.0867947
#>  [2,] 2.3903865 0.04103371 1.9718864
#>  [3,] 3.4045783 0.71152931 4.2575134
#>  [4,] 0.2837894 0.79791951 3.4257135
#>  [5,] 0.8942045 3.99104628 0.1432913
#>  [6,] 0.5894129 0.48498281 0.1796983
#>  [7,] 0.2862989 0.23350954 1.3905292
#>  [8,] 1.9559629 0.91220838 0.4021000
#>  [9,] 0.8150449 1.85256862 0.8325032
#> [10,] 0.4103970 0.59558436 0.5207148

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32425385 1.10572002 0.30262848
#>  [2,] 0.98312602 0.06285358 0.06285358
#>  [3,] 0.33071407 0.05678894 0.33071407
#>  [4,] 0.59516076 0.94169798 0.94169798
#>  [5,] 0.58459568 0.60052124 0.54352862
#>  [6,] 1.17491865 0.85082037 1.22648404
#>  [7,] 1.46486673 0.21515217 0.73893281
#>  [8,] 0.38875080 0.38875080 0.07077157
#>  [9,] 0.56515128 0.56515128 0.56515128
#> [10,] 0.07646599 0.28301438 1.37430760
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14812775 0.15404230 0.14812775
#>  [2,] 0.62455647 3.03589265 0.62455647
#>  [3,] 0.19534926 1.61954596 1.82827685
#>  [4,] 0.39988135 2.73480246 1.34815140
#>  [5,] 0.63903839 0.63903839 0.63903839
#>  [6,] 0.02961493 0.02961493 0.02961493
#>  [7,] 0.34454845 0.08900914 0.17944309
#>  [8,] 0.21509177 0.11885321 0.80828082
#>  [9,] 0.98213137 0.33598002 0.98213137
#> [10,] 0.43071611 1.04735598 0.17314125
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.03965251 0.12745481 0.1274548
#>  [2,] 0.03198685 0.03198685 0.4967036
#>  [3,] 0.93202155 0.05385679 0.8223276
#>  [4,] 0.35298659 0.81783892 1.2422758
#>  [5,] 0.33635978 0.33635978 0.2478904
#>  [6,] 0.13430419 1.52155913 0.7487808
#>  [7,] 1.10278915 1.10278915 1.0625233
#>  [8,] 0.41050427 0.37521477 0.3985576
#>  [9,] 0.38924395 0.25511299 0.1705865
#> [10,] 0.29410346 0.09450516 1.3056054
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.97509100 0.34763514 0.632201659
#>  [2,] 0.46720176 0.45488896 0.520323247
#>  [3,] 0.25404214 0.25404214 0.002234728
#>  [4,] 1.00487371 0.19895013 0.198950129
#>  [5,] 1.37266723 0.11311581 0.113115808
#>  [6,] 0.03898398 1.78547572 2.063935837
#>  [7,] 0.28300528 0.28300528 0.283005280
#>  [8,] 0.14194425 0.14194425 0.141944253
#>  [9,] 1.07882627 1.07882627 4.108288876
#> [10,] 0.06473435 0.06473435 0.322108331
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.14724336 0.1472434 0.14724336
#>  [2,] 0.38408985 0.2528380 2.08661220
#>  [3,] 0.43796702 0.4379670 0.43796702
#>  [4,] 0.20101610 0.6582857 0.35292833
#>  [5,] 0.25158963 0.2515896 0.25158963
#>  [6,] 0.68015615 0.6801561 0.36348927
#>  [7,] 0.06627237 0.3179237 0.98118571
#>  [8,] 2.16529596 0.5638845 0.08963783
#>  [9,] 0.53445333 0.5623548 1.07842776
#> [10,] 1.04284671 1.7636369 0.13851450

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 8.3949354 8.2810317 0.4106790
#>  [2,] 8.1621214 3.9442568 1.9615002
#>  [3,] 3.1241045 3.1634299 0.6378620
#>  [4,] 4.0080933 1.2473698 0.5084819
#>  [5,] 2.1474031 2.1127814 0.7710406
#>  [6,] 0.1840655 0.3290720 0.3290720
#>  [7,] 0.8367397 0.5768739 1.6152630
#>  [8,] 1.1755847 1.1755847 0.8345499
#>  [9,] 0.4688111 0.8840505 3.7855629
#> [10,] 1.5425811 1.8950451 1.2825344
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11380994 0.44404876 0.66803197
#>  [2,] 1.09528507 0.52475232 3.61736981
#>  [3,] 0.06930516 0.06930516 0.06930516
#>  [4,] 0.51825114 7.68621070 1.52466724
#>  [5,] 2.54205813 2.19905587 0.24513468
#>  [6,] 0.56810883 0.68802222 0.68802222
#>  [7,] 0.33474003 0.33474003 0.01710978
#>  [8,] 0.83101585 1.00064092 0.09307200
#>  [9,] 0.39355229 0.51652528 1.85211379
#> [10,] 1.57667889 2.52833839 2.52833839
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9318733 0.93187326 0.9318733
#>  [2,] 0.1761434 0.03395908 1.1765695
#>  [3,] 1.1935928 7.03600963 1.0618478
#>  [4,] 2.2669049 1.68935358 0.7155424
#>  [5,] 0.8486382 1.59332096 1.5933210
#>  [6,] 0.2646660 0.26466596 0.2646660
#>  [7,] 1.1096803 0.39996712 0.4678439
#>  [8,] 0.2826049 0.66686722 1.6398328
#>  [9,] 0.2969026 2.18093369 0.5430138
#> [10,] 2.0664490 2.42462680 0.6017626
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 1.36396891 1.36396891 1.363968908
#>  [2,] 0.62344968 0.62344968 0.623449679
#>  [3,] 0.03862261 0.05626852 0.127533745
#>  [4,] 0.87163348 0.03207286 0.871633476
#>  [5,] 0.46161837 3.74972073 0.602299352
#>  [6,] 1.67910907 1.67910907 1.679109070
#>  [7,] 0.14953333 0.01979879 1.249739113
#>  [8,] 1.46964406 0.93700311 1.245872465
#>  [9,] 1.23566611 0.64925002 0.003049959
#> [10,] 0.20663499 0.20663499 0.206634989
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.22829325 0.2695221 0.2695221
#>  [2,] 2.91457884 1.4578350 1.2273643
#>  [3,] 0.36820072 0.3682007 0.3682007
#>  [4,] 0.45895687 1.0204124 5.0825302
#>  [5,] 0.09965621 1.0189247 0.1022602
#>  [6,] 1.61757143 0.5013242 1.0020860
#>  [7,] 1.93262952 0.3539551 0.4757971
#>  [8,] 1.79084145 1.3169676 0.6399313
#>  [9,] 1.82381495 1.8238149 0.1079170
#> [10,] 3.94506148 0.6858361 3.6418054

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40184112 0.40184112 0.40184112
#>  [2,] 0.25971255 1.05707390 0.02080736
#>  [3,] 0.10478782 0.24989467 0.24989467
#>  [4,] 0.24662846 0.47309481 0.24662846
#>  [5,] 0.34225316 0.34225316 0.34225316
#>  [6,] 0.26736254 0.26736254 0.26736254
#>  [7,] 0.35987316 2.15031525 0.63921137
#>  [8,] 0.09044728 0.09044728 0.09044728
#>  [9,] 0.66298751 0.72535347 1.47426789
#> [10,] 0.32852713 0.32852713 0.32852713
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17265510 0.17265510 0.17265510
#>  [2,] 0.21862762 0.41460142 0.21862762
#>  [3,] 0.77538544 0.77538544 0.23435577
#>  [4,] 0.70589828 0.78429033 0.78429033
#>  [5,] 0.05603638 0.05603638 0.05603638
#>  [6,] 0.04927649 0.04927649 0.04927649
#>  [7,] 0.98057884 0.14210981 0.30101913
#>  [8,] 1.06920805 1.06920805 1.06920805
#>  [9,] 2.67458874 0.09216003 0.45948707
#> [10,] 0.08824681 0.08824681 0.08824681
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.06065499 0.7581270 0.08706872
#>  [2,] 0.21702195 0.2210035 0.21702195
#>  [3,] 0.13540431 0.1354043 0.71848925
#>  [4,] 0.93203942 2.9957315 2.96251533
#>  [5,] 1.23710573 1.0453350 1.30258392
#>  [6,] 1.04877892 0.4872686 0.92063610
#>  [7,] 0.56073700 0.4996572 0.56073700
#>  [8,] 0.21781372 0.2178137 0.44363321
#>  [9,] 0.79095545 1.4327483 0.41985178
#> [10,] 0.40393468 1.2272227 1.29568387
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14200284 0.14200284 0.14200284
#>  [2,] 0.79529573 1.19045144 1.00098745
#>  [3,] 0.03655298 0.03655298 0.03655298
#>  [4,] 1.05212836 0.09784685 0.79922057
#>  [5,] 1.25456649 1.18676690 1.67533727
#>  [6,] 0.07112309 0.20405770 0.20405770
#>  [7,] 0.76952233 0.32671388 2.86028307
#>  [8,] 0.18374278 0.18374278 0.18374278
#>  [9,] 0.14624975 0.14624975 0.14624975
#> [10,] 0.30135911 0.95562270 0.30135911
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06770234 0.17551437 0.50390020
#>  [2,] 0.10456155 0.10456155 0.10456155
#>  [3,] 0.82068046 0.18499396 0.25564241
#>  [4,] 0.32001355 0.32001355 0.32001355
#>  [5,] 0.06531435 0.06531435 0.06531435
#>  [6,] 0.89744475 0.89744475 0.89744475
#>  [7,] 0.68583654 0.24617760 0.12795203
#>  [8,] 0.03854352 0.04082803 0.72469390
#>  [9,] 0.07133597 0.07133597 0.07133597
#> [10,] 0.57003533 1.01893340 0.19415167

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13479940 0.13479940 0.13479940
#>  [2,] 0.22783664 0.50734901 0.02738458
#>  [3,] 0.49455280 0.40587922 0.31633953
#>  [4,] 0.04988792 0.22855025 0.23999777
#>  [5,] 0.06614717 0.11876007 0.06614717
#>  [6,] 0.73581960 0.76895098 0.18842939
#>  [7,] 0.48867370 0.06730138 0.06730138
#>  [8,] 0.80618645 0.67719811 0.67719811
#>  [9,] 0.45693327 1.07172994 1.67749946
#> [10,] 0.01453316 0.18609172 0.34806215
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.18844839 0.07404689 0.16747032
#>  [2,] 0.11996789 0.32443255 0.28428173
#>  [3,] 0.14080928 0.17228721 0.17228721
#>  [4,] 0.33347007 0.55425611 0.35672049
#>  [5,] 0.52875452 0.13348396 0.06953088
#>  [6,] 0.83645383 0.83645383 0.83645383
#>  [7,] 0.01584686 0.29064847 0.29064847
#>  [8,] 1.02596398 1.02596398 0.32667210
#>  [9,] 0.07436859 0.07436859 0.17601784
#> [10,] 0.22681659 0.22681659 0.16273444
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.53315878 0.53315878 0.53315878
#>  [2,] 0.26372138 0.51810731 0.15460778
#>  [3,] 0.68023258 0.41552672 0.66296494
#>  [4,] 0.08616748 0.08310586 0.08616748
#>  [5,] 0.52308750 0.17152038 0.52308750
#>  [6,] 0.18905914 0.31271629 0.31271629
#>  [7,] 0.32189446 1.18541195 0.41447658
#>  [8,] 1.14649587 0.21738038 0.32440408
#>  [9,] 0.49159707 0.49042567 0.49042567
#> [10,] 0.02047786 0.53788519 0.30404198
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.20465344 0.78115893 0.20465344
#>  [2,] 0.89021205 0.68734121 1.12583509
#>  [3,] 0.26066024 1.22165866 1.22165866
#>  [4,] 0.30700178 0.32799357 0.21956136
#>  [5,] 0.16460163 0.16460163 0.16460163
#>  [6,] 0.18936409 0.16124485 0.03550814
#>  [7,] 0.09155994 0.09246482 0.12842629
#>  [8,] 0.06142368 0.01175137 0.06142368
#>  [9,] 0.07083692 0.12619259 0.12619259
#> [10,] 0.72776643 0.72776643 0.72776643

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.284769508 0.28476951 0.28476951
#>  [2,] 0.448674764 0.43842832 0.44867476
#>  [3,] 0.137111320 0.13711132 0.41832804
#>  [4,] 0.004835353 0.47993489 0.47993489
#>  [5,] 0.403576195 0.03094507 0.60102766
#>  [6,] 0.369872997 0.19581210 0.36987300
#>  [7,] 0.270701428 0.96075698 0.05020613
#>  [8,] 0.128562480 0.14549302 0.17409822
#>  [9,] 0.329674509 0.24995949 1.42700474
#> [10,] 0.416806490 0.11190026 0.41680649
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5034004 0.10302936 0.73919340
#>  [2,] 0.5857498 0.15355371 1.45764540
#>  [3,] 0.1242197 0.06801976 0.47486117
#>  [4,] 1.0827657 0.42513792 1.08276565
#>  [5,] 0.1069823 0.24199319 0.03913803
#>  [6,] 0.1727200 0.54331172 0.46328763
#>  [7,] 0.2980996 0.58730340 0.30342476
#>  [8,] 0.4546641 0.45466414 0.16067749
#>  [9,] 0.1335704 0.13570042 0.24672175
#> [10,] 0.2827578 0.28275777 0.28275777
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38017229 0.50078977 0.69803796
#>  [2,] 1.13233083 0.07914742 0.53739054
#>  [3,] 0.04707099 0.21592450 0.17616877
#>  [4,] 0.13292906 0.15523534 0.15523534
#>  [5,] 0.03953914 0.03953914 0.03953914
#>  [6,] 0.05375769 0.05375769 0.05375769
#>  [7,] 0.05047122 0.22545314 0.05047122
#>  [8,] 0.06511684 0.46273773 0.20629166
#>  [9,] 0.93266491 1.28930812 0.11797450
#> [10,] 0.14905307 0.02477101 0.20634430
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.59371155 0.73604026 0.59371155
#>  [2,] 0.19738663 0.26317362 0.11472535
#>  [3,] 0.51538343 1.26862661 0.00851354
#>  [4,] 0.05814344 0.05814344 0.05814344
#>  [5,] 0.57442598 0.41103781 0.97531491
#>  [6,] 0.27553546 0.21926105 0.75175936
#>  [7,] 0.37169179 0.37169179 0.37169179
#>  [8,] 0.94405777 0.94405777 0.80316084
#>  [9,] 0.36784138 0.36784138 0.36784138
#> [10,] 0.12775603 0.12775603 0.12775603

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3206174 0.32061744 0.32061744
#>  [2,] 0.2498767 0.49720225 0.24987667
#>  [3,] 0.2673612 0.26736117 0.65373024
#>  [4,] 0.5048889 0.13305336 0.15236908
#>  [5,] 0.2518067 0.44071859 0.44071859
#>  [6,] 0.5316932 0.08445708 0.08445708
#>  [7,] 0.3298281 0.08543230 0.32982809
#>  [8,] 0.3427717 0.32604687 0.27501649
#>  [9,] 0.1536669 0.15366694 0.15366694
#> [10,] 0.9658791 0.96587908 0.61987105
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.19127520 0.92928940 0.19127520
#>  [2,] 1.10272284 1.37278626 0.04277637
#>  [3,] 1.69823340 0.02101255 0.53168734
#>  [4,] 0.17199176 0.20124925 0.20124925
#>  [5,] 0.72090118 0.16166320 0.73639445
#>  [6,] 0.05262329 0.10747394 0.26322331
#>  [7,] 0.52689164 0.51293508 0.52689164
#>  [8,] 0.69816170 0.69816170 0.10965356
#>  [9,] 0.71201346 0.82714312 1.10334727
#> [10,] 0.23803381 0.16076754 0.48089697
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36141659 0.05714972 0.71086208
#>  [2,] 0.26572952 0.40128567 0.03017004
#>  [3,] 0.41567115 0.37396836 0.41567115
#>  [4,] 0.09790281 0.14846389 0.09790281
#>  [5,] 0.31666421 0.31666421 0.09403060
#>  [6,] 1.66036575 1.66036575 0.58858296
#>  [7,] 0.23842115 0.18280892 0.23842115
#>  [8,] 0.07863880 0.08712424 0.27147491
#>  [9,] 0.32567337 0.32567337 0.32567337
#> [10,] 1.49188004 0.52488182 0.52488182
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32463986 0.11347057 0.21945352
#>  [2,] 0.64834337 0.39887297 0.64834337
#>  [3,] 0.22237715 0.22237715 0.30352407
#>  [4,] 0.13728601 0.57517452 0.53157984
#>  [5,] 0.12476259 0.12476259 0.12476259
#>  [6,] 0.94053626 0.16255766 0.93253374
#>  [7,] 0.05728703 0.05728703 0.05728703
#>  [8,] 0.60824361 0.60824361 0.09720236
#>  [9,] 0.01930821 0.01930821 0.01930821
#> [10,] 0.02996509 0.02996509 0.14496820