Sample from parametrized extendible MO distributions

Draws n iid samples from a d-variate paramatrized extendible MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameters, see Details.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with n rows and d columns with the rows corresponding to iid distributed samples of a d-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family

We have \(\nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

Exchangeable Marshall–Olkin simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

Lévy-frailty simulation algorithms

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9387302 2.57176905 1.8507609
#>  [2,] 0.4970415 0.49704147 0.4970415
#>  [3,] 4.8187065 1.54165284 1.1057476
#>  [4,] 0.2367109 0.06832267 1.0284184
#>  [5,] 1.0337250 6.12880120 3.4935591
#>  [6,] 1.4501241 5.05414061 1.6419297
#>  [7,] 0.2879831 3.54186130 2.3309389
#>  [8,] 4.2264552 5.84806080 0.9852624
#>  [9,] 0.9863754 0.75603783 1.1002419
#> [10,] 0.8579073 2.25792673 0.1442562
## comonotone
rpextmo(10, 3, a = 1)
#>             [,1]       [,2]       [,3]
#>  [1,] 4.50612121 4.50612121 4.50612121
#>  [2,] 1.93352277 1.93352277 1.93352277
#>  [3,] 2.42425812 2.42425812 2.42425812
#>  [4,] 0.09541686 0.09541686 0.09541686
#>  [5,] 0.94802249 0.94802249 0.94802249
#>  [6,] 0.45901359 0.45901359 0.45901359
#>  [7,] 1.19224084 1.19224084 1.19224084
#>  [8,] 1.53621527 1.53621527 1.53621527
#>  [9,] 0.23956826 0.23956826 0.23956826
#> [10,] 1.11208536 1.11208536 1.11208536
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.18863155 0.03615898 0.2237412
#>  [2,] 0.55558971 0.47571315 1.4669308
#>  [3,] 0.01910137 0.06931077 1.0490142
#>  [4,] 0.15593392 0.25170913 0.4956804
#>  [5,] 1.94357654 4.39896728 0.6802749
#>  [6,] 1.35312717 0.62122770 1.9162499
#>  [7,] 0.49235817 1.19030782 2.7489923
#>  [8,] 2.60468999 0.74902607 0.2576092
#>  [9,] 0.53799220 0.37950259 1.2924421
#> [10,] 0.25455231 0.93246184 0.4463176

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.01150934 4.18653715 2.8454737
#>  [2,] 1.66275422 2.28670391 0.1664680
#>  [3,] 0.44895226 1.34064826 5.6594023
#>  [4,] 0.66863493 0.45134481 0.6686349
#>  [5,] 0.30191644 0.27774080 4.2453506
#>  [6,] 1.02662171 0.07773814 0.9221033
#>  [7,] 3.05408037 0.82816663 0.6068075
#>  [8,] 0.03623071 2.55101392 0.3418486
#>  [9,] 0.36202571 2.97526778 2.9752678
#> [10,] 0.80035762 1.91475104 0.8278428
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2921870 0.2921870 0.2921870
#>  [2,] 0.2756261 0.2756261 0.2756261
#>  [3,] 1.8896550 1.8896550 1.8896550
#>  [4,] 0.6327416 0.6327416 0.6327416
#>  [5,] 2.0692155 2.0692155 2.0692155
#>  [6,] 0.3262571 0.3262571 0.3262571
#>  [7,] 0.6347321 0.6347321 0.6347321
#>  [8,] 1.4727558 1.4727558 1.4727558
#>  [9,] 0.7801704 0.7801704 0.7801704
#> [10,] 0.9266221 0.9266221 0.9266221
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 4.15759595 2.1840381 0.2640941
#>  [2,] 0.15871330 0.7021799 0.7956672
#>  [3,] 0.75994654 0.2878108 1.3497126
#>  [4,] 0.52027151 1.0394646 0.5737331
#>  [5,] 1.04813608 0.7750283 1.4219898
#>  [6,] 0.35554708 0.7543779 1.7240723
#>  [7,] 1.24615605 1.0037212 0.3597527
#>  [8,] 0.67898817 0.3534365 0.1590034
#>  [9,] 0.02039482 0.1826208 0.5099151
#> [10,] 1.19758795 0.6501104 0.4289677

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7014144 1.13420359 1.1342036
#>  [2,] 4.2986142 0.36077498 1.2006161
#>  [3,] 1.7024292 1.37543446 1.7024292
#>  [4,] 2.1002042 0.28951869 0.5254545
#>  [5,] 0.4944705 0.20553491 0.6984827
#>  [6,] 2.0115537 1.49008957 0.3097904
#>  [7,] 0.5353819 0.05700866 0.8425456
#>  [8,] 0.3370524 0.10350657 1.7154394
#>  [9,] 1.8033688 2.59182340 3.4013633
#> [10,] 0.5711427 0.31665802 0.5366644
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34659237 0.34659237 0.34659237
#>  [2,] 0.01190017 0.01190017 0.01190017
#>  [3,] 0.84628370 0.84628370 0.84628370
#>  [4,] 0.91307697 0.91307697 0.91307697
#>  [5,] 0.95091934 0.95091934 0.95091934
#>  [6,] 2.59460836 2.59460836 2.59460836
#>  [7,] 0.30104212 0.30104212 0.30104212
#>  [8,] 0.77408975 0.77408975 0.77408975
#>  [9,] 1.29796424 1.29796424 1.29796424
#> [10,] 1.00756307 1.00756307 1.00756307
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>            [,1]       [,2]       [,3]
#>  [1,] 1.0714249 2.16124929 0.06245639
#>  [2,] 1.1743722 1.41806395 0.47385178
#>  [3,] 0.6777618 0.42520862 2.78224402
#>  [4,] 0.3477036 2.52843727 0.15300394
#>  [5,] 2.0017847 0.78645377 1.42997357
#>  [6,] 0.1304996 0.21722126 0.28744116
#>  [7,] 1.0512698 0.39139502 0.85469516
#>  [8,] 0.2638505 1.61623963 0.91122819
#>  [9,] 0.4117026 0.14591273 1.65593503
#> [10,] 0.2227447 0.02859994 1.48497543

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.44142517 0.12975936 0.44142517
#>  [2,] 7.86703054 1.52660488 2.32517362
#>  [3,] 0.05111554 0.04604049 1.96836449
#>  [4,] 1.35084121 0.99107807 1.35084121
#>  [5,] 0.04485732 2.56958371 0.81127310
#>  [6,] 0.19105456 0.32502432 2.26293087
#>  [7,] 4.99323168 0.54315786 1.36785943
#>  [8,] 4.08582831 0.95258466 3.77573973
#>  [9,] 0.00191538 0.00191538 0.00191538
#> [10,] 0.09009717 0.49711686 0.42142473
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08480294 0.08480294 0.08480294
#>  [2,] 4.08410737 4.08410737 4.08410737
#>  [3,] 0.57369494 0.57369494 0.57369494
#>  [4,] 2.46475054 2.46475054 2.46475054
#>  [5,] 0.46151525 0.46151525 0.46151525
#>  [6,] 0.07381092 0.07381092 0.07381092
#>  [7,] 0.84999860 0.84999860 0.84999860
#>  [8,] 0.70570388 0.70570388 0.70570388
#>  [9,] 0.50111500 0.50111500 0.50111500
#> [10,] 0.15569163 0.15569163 0.15569163
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>            [,1]      [,2]       [,3]
#>  [1,] 4.0903985 1.7746166 0.28842721
#>  [2,] 0.2318448 0.9626988 0.11254017
#>  [3,] 0.5955988 0.5096592 2.38006749
#>  [4,] 0.8232776 2.2667041 0.22222859
#>  [5,] 0.5359287 1.3220249 1.61512831
#>  [6,] 0.7956634 0.1551256 1.10506270
#>  [7,] 0.1240952 1.3328285 3.55063440
#>  [8,] 1.6154421 2.6976921 0.75728618
#>  [9,] 1.4874954 1.5078029 0.70356501
#> [10,] 0.7137028 1.1319874 0.02193812

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.8908777 0.9037603 2.4094709
#>  [2,] 1.2838753 1.8057043 0.5067451
#>  [3,] 0.6732719 2.6733685 0.4102813
#>  [4,] 0.4212726 1.0294178 0.5464502
#>  [5,] 4.7839104 0.3387259 4.7839104
#>  [6,] 0.9066169 2.1117348 0.5050414
#>  [7,] 0.6983780 7.0443212 3.3891895
#>  [8,] 2.0664991 0.9751298 3.3854565
#>  [9,] 0.4916294 0.4916294 0.4916294
#> [10,] 0.7563439 0.6959187 1.4456366
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 3.88637803 3.88637803 3.88637803
#>  [2,] 0.85558752 0.85558752 0.85558752
#>  [3,] 0.11400864 0.11400864 0.11400864
#>  [4,] 0.04721669 0.04721669 0.04721669
#>  [5,] 0.20321540 0.20321540 0.20321540
#>  [6,] 0.96243207 0.96243207 0.96243207
#>  [7,] 0.53041850 0.53041850 0.53041850
#>  [8,] 1.63996172 1.63996172 1.63996172
#>  [9,] 0.93846487 0.93846487 0.93846487
#> [10,] 0.23624701 0.23624701 0.23624701
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>              [,1]        [,2]      [,3]
#>  [1,] 4.178693392 0.239755166 0.7881172
#>  [2,] 0.008487075 0.944684474 0.1101266
#>  [3,] 0.065292579 0.075850802 2.3456796
#>  [4,] 0.411777194 0.040444640 0.9154839
#>  [5,] 1.655087435 0.609656062 0.9199193
#>  [6,] 1.937063301 0.003258475 0.9606196
#>  [7,] 0.456398542 0.348352248 1.8847903
#>  [8,] 0.301398922 2.829639327 1.2564102
#>  [9,] 0.510379705 0.348945639 0.9095684
#> [10,] 1.093342915 0.354145045 3.8922550

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.93106503 0.87587972 0.93106503
#>  [2,] 1.95722224 1.00035069 0.51236050
#>  [3,] 2.31224346 1.77243953 0.84244035
#>  [4,] 2.07314216 1.65358587 0.29529819
#>  [5,] 0.50881597 0.40407364 1.11580892
#>  [6,] 0.14415053 0.14415053 0.14415053
#>  [7,] 0.03014978 0.03014978 0.03014978
#>  [8,] 2.95375930 8.17210979 1.10715747
#>  [9,] 0.85783950 0.88832443 0.95190288
#> [10,] 0.08890594 5.16287761 1.94081963
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.5976265 0.5976265 0.5976265
#>  [2,] 0.8563056 0.8563056 0.8563056
#>  [3,] 0.6497184 0.6497184 0.6497184
#>  [4,] 0.9068048 0.9068048 0.9068048
#>  [5,] 1.6822739 1.6822739 1.6822739
#>  [6,] 4.0919218 4.0919218 4.0919218
#>  [7,] 0.7952500 0.7952500 0.7952500
#>  [8,] 2.5270249 2.5270249 2.5270249
#>  [9,] 0.2000703 0.2000703 0.2000703
#> [10,] 0.1500828 0.1500828 0.1500828
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>            [,1]        [,2]       [,3]
#>  [1,] 0.2630273 0.196275313 0.23530050
#>  [2,] 1.5317777 1.689706983 0.28928202
#>  [3,] 1.5324238 2.248748893 0.66594211
#>  [4,] 0.1235717 0.043351731 0.66587020
#>  [5,] 0.1592813 1.922657261 1.00155160
#>  [6,] 0.9825633 0.905007526 2.47826513
#>  [7,] 0.3940454 0.260103367 2.56610011
#>  [8,] 0.4561724 0.375000057 2.32806449
#>  [9,] 1.4491474 0.344895941 0.10114377
#> [10,] 0.6132083 0.002268274 0.01185685

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9401966 0.7148115 0.9401966
#>  [2,] 2.8980760 1.3825871 2.5394666
#>  [3,] 4.7992262 0.1111968 1.6199174
#>  [4,] 2.9941331 1.6245411 0.2384354
#>  [5,] 3.3234908 0.3944563 4.4665410
#>  [6,] 1.6758341 1.6758341 1.1978162
#>  [7,] 0.4756780 0.4756780 0.4756780
#>  [8,] 1.4207541 1.4777511 2.9197583
#>  [9,] 0.7326715 0.5292544 4.3262478
#> [10,] 3.6035337 1.3023472 0.2153714
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "ESM"
)
#>               [,1]         [,2]         [,3]
#>  [1,] 0.0008280394 0.0008280394 0.0008280394
#>  [2,] 0.4125790303 0.4125790303 0.4125790303
#>  [3,] 2.5937403607 2.5937403607 2.5937403607
#>  [4,] 0.2377665434 0.2377665434 0.2377665434
#>  [5,] 0.6021442846 0.6021442846 0.6021442846
#>  [6,] 0.1492228612 0.1492228612 0.1492228612
#>  [7,] 0.4943139376 0.4943139376 0.4943139376
#>  [8,] 0.1500877765 0.1500877765 0.1500877765
#>  [9,] 0.4382753870 0.4382753870 0.4382753870
#> [10,] 0.7400305084 0.7400305084 0.7400305084
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.99032185 0.7027152 1.0211805
#>  [2,] 1.08110815 1.5247945 3.0315402
#>  [3,] 0.01072056 1.1405908 2.7109947
#>  [4,] 2.52450052 1.0646155 0.9453594
#>  [5,] 1.59775426 2.8903608 0.1237210
#>  [6,] 1.03050871 0.3599957 0.4195047
#>  [7,] 0.43093590 1.8708427 1.5726718
#>  [8,] 0.28898572 1.9749357 1.1856840
#>  [9,] 2.25355670 0.6930438 7.8176636
#> [10,] 1.92784855 1.1425821 0.9503838

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.4212584 1.1597065 0.2209875
#>  [2,] 0.6294976 1.4982526 0.6098567
#>  [3,] 0.7464464 1.7002670 1.4311225
#>  [4,] 1.3394013 0.2202860 2.7074685
#>  [5,] 2.4967068 1.1401622 0.4288468
#>  [6,] 2.2899630 1.4561915 2.2899630
#>  [7,] 3.4089937 0.1307988 0.8765990
#>  [8,] 0.9291411 0.9291411 0.2907726
#>  [9,] 2.1241549 1.3745911 1.7607411
#> [10,] 1.5044824 2.2606187 2.0903361
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3642267 0.3642267 0.3642267
#>  [2,] 1.8284538 1.8284538 1.8284538
#>  [3,] 2.0065692 2.0065692 2.0065692
#>  [4,] 1.9430641 1.9430641 1.9430641
#>  [5,] 2.2484413 2.2484413 2.2484413
#>  [6,] 1.5166018 1.5166018 1.5166018
#>  [7,] 2.4277599 2.4277599 2.4277599
#>  [8,] 0.4181192 0.4181192 0.4181192
#>  [9,] 1.3044519 1.3044519 1.3044519
#> [10,] 1.4085973 1.4085973 1.4085973
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8427077 0.8486622 0.3200242
#>  [2,] 1.2618036 0.7336840 0.3144376
#>  [3,] 0.1404810 0.9746551 1.3523196
#>  [4,] 0.5080712 0.6535140 0.4285994
#>  [5,] 0.9828397 1.2911084 1.2400992
#>  [6,] 1.3666574 0.2646425 5.0925437
#>  [7,] 2.2895803 0.3106342 0.9428243
#>  [8,] 0.9143969 2.6514594 0.6323722
#>  [9,] 0.3425485 0.1133844 3.6359353
#> [10,] 1.5700428 1.8627147 1.2317471

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 3.09731498 1.92952118 3.0973150
#>  [2,] 1.15790203 0.13290376 1.9960095
#>  [3,] 0.63018103 0.06305931 0.3383326
#>  [4,] 1.11638310 0.45479280 0.4967439
#>  [5,] 0.26924836 4.84530521 1.7420737
#>  [6,] 0.08605508 2.89050194 4.3965360
#>  [7,] 0.90544220 0.74801323 0.5909158
#>  [8,] 1.31425030 1.26819981 1.3142503
#>  [9,] 0.76041297 0.24298033 0.9506319
#> [10,] 1.17897447 0.31564933 2.7984148
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.57624600 0.57624600 0.57624600
#>  [2,] 2.46662679 2.46662679 2.46662679
#>  [3,] 1.06962405 1.06962405 1.06962405
#>  [4,] 0.09698185 0.09698185 0.09698185
#>  [5,] 0.73648798 0.73648798 0.73648798
#>  [6,] 0.09242687 0.09242687 0.09242687
#>  [7,] 1.14441689 1.14441689 1.14441689
#>  [8,] 1.02589623 1.02589623 1.02589623
#>  [9,] 3.30237507 3.30237507 3.30237507
#> [10,] 0.30000450 0.30000450 0.30000450
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "MDCM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.16389822 0.5855730 0.03958048
#>  [2,] 0.11120857 0.3263699 0.12411323
#>  [3,] 3.99643423 0.9671963 1.50741853
#>  [4,] 1.72775463 0.3089088 0.33244346
#>  [5,] 0.09575745 0.4223514 1.30986508
#>  [6,] 0.49581687 2.1474846 1.53286548
#>  [7,] 0.18870379 0.5948257 0.24375015
#>  [8,] 0.48119963 2.1255151 2.40478935
#>  [9,] 3.30821077 0.5511228 0.15271091
#> [10,] 1.32785730 0.2799659 1.25609952

rpextmo(
    10, 3,
    a = 0.2, b = 0.5,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.25645457 1.17117857 3.35544528
#>  [2,] 1.06794697 1.06794697 1.06794697
#>  [3,] 0.89374463 1.40322163 1.40322163
#>  [4,] 0.26103580 0.26103580 0.26103580
#>  [5,] 0.71859197 0.62165150 3.89423901
#>  [6,] 0.08649624 0.08649624 0.08649624
#>  [7,] 0.35168902 0.35168902 0.35168902
#>  [8,] 0.29890240 0.29890240 0.29890240
#>  [9,] 0.16430712 1.93675275 0.23022627
#> [10,] 0.13924895 0.20357050 2.45915245
## comonotone
rpextmo(
    10, 3,
    a = 1,
    family = "Armageddon",
    method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7028032 0.7028032 0.7028032
#>  [2,] 1.4240266 1.4240266 1.4240266
#>  [3,] 3.4946978 3.4946978 3.4946978
#>  [4,] 2.2557155 2.2557155 2.2557155
#>  [5,] 0.9000986 0.9000986 0.9000986
#>  [6,] 0.0111608 0.0111608 0.0111608
#>  [7,] 0.6503234 0.6503234 0.6503234
#>  [8,] 0.2365831 0.2365831 0.2365831
#>  [9,] 0.3503832 0.3503832 0.3503832
#> [10,] 0.6175198 0.6175198 0.6175198
## independence
rpextmo(
    10, 3,
    b = 1,
    family = "Armageddon",
    method = "AM"
)
#>             [,1]      [,2]       [,3]
#>  [1,] 0.19813862 0.2343487 0.83035015
#>  [2,] 0.83935345 0.6011946 0.93092538
#>  [3,] 2.36507061 0.2327009 0.25756946
#>  [4,] 0.86218876 0.5101125 1.49342769
#>  [5,] 0.42025469 0.1434152 0.06911597
#>  [6,] 0.10566201 0.2510737 1.87955682
#>  [7,] 0.01548297 1.1438603 0.45977081
#>  [8,] 1.17390255 0.3524189 0.37580066
#>  [9,] 1.29925780 0.9228249 1.21704532
#> [10,] 1.09070810 0.1844003 0.11391844

## Poisson
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03790535 0.03790535 0.18290395
#>  [2,] 0.10871372 0.32829956 0.10871372
#>  [3,] 1.14118382 0.15435840 0.06358779
#>  [4,] 0.20596735 0.20596735 1.68217846
#>  [5,] 0.03151825 0.56487840 0.03151825
#>  [6,] 0.32251956 0.16378921 0.66913740
#>  [7,] 0.04375132 0.27428345 0.23786479
#>  [8,] 0.13072123 0.00758630 0.13072123
#>  [9,] 0.49997419 0.59435143 0.85484896
#> [10,] 0.19006911 1.01923283 0.19006911
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.77875811 0.3022610 0.7787581
#>  [2,] 0.51397131 0.2652324 0.2268991
#>  [3,] 0.88836829 1.2167707 0.3174749
#>  [4,] 0.94389713 0.3947427 3.9163856
#>  [5,] 0.33635783 0.1758587 0.4592758
#>  [6,] 0.32888374 0.3288837 0.0225250
#>  [7,] 0.50885157 0.1126562 0.4401287
#>  [8,] 0.30100763 0.6426618 0.9357586
#>  [9,] 0.55795371 0.6068626 0.6068626
#> [10,] 0.04232404 0.2533435 0.8331086
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2006806 0.5581613 0.1325488
#>  [2,] 0.2681347 0.2681347 0.3771432
#>  [3,] 0.2915479 0.9293930 2.1320527
#>  [4,] 0.2166793 0.5155152 0.5155152
#>  [5,] 0.1824681 0.2835562 1.1920650
#>  [6,] 0.1567817 0.1567817 0.1146660
#>  [7,] 0.4104546 1.2253513 0.4104546
#>  [8,] 1.4463546 0.1011086 1.4463546
#>  [9,] 1.4453152 0.6214032 1.4453152
#> [10,] 1.5014303 0.4820175 0.5452508
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "MDCM"
)
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8241565 0.8241565 1.67405051
#>  [2,] 0.6574123 0.6574123 0.01319648
#>  [3,] 1.1323618 0.2234734 0.14457240
#>  [4,] 1.3641614 0.9567463 0.95674632
#>  [5,] 0.6493197 0.6493197 0.29458704
#>  [6,] 0.1166990 0.1166990 0.11669898
#>  [7,] 1.1979489 2.0384491 0.27761611
#>  [8,] 0.9676728 0.3846498 0.31478860
#>  [9,] 0.4285371 0.5497061 0.11379670
#> [10,] 0.6142190 1.0878721 0.07142166
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Poisson",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.31896874 1.69219301 0.31896874
#>  [2,] 0.36108063 0.36108063 0.36108063
#>  [3,] 0.48132187 0.72682966 0.65574090
#>  [4,] 0.05136044 0.05136044 0.05136044
#>  [5,] 0.69557451 0.27652618 0.18083528
#>  [6,] 0.02939202 0.48760545 0.90657942
#>  [7,] 0.38724253 0.38724253 1.85666581
#>  [8,] 0.26077729 0.09607896 0.46599857
#>  [9,] 1.08790138 0.39641371 1.23299105
#> [10,] 0.27290046 1.35954147 0.27290046

## Pareto
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36856601 2.31127580 2.29017955
#>  [2,] 0.61862979 2.23169644 0.41765532
#>  [3,] 1.27814082 0.62313975 0.57054430
#>  [4,] 1.21900395 0.22271796 1.21900395
#>  [5,] 1.63224487 0.21278474 0.04934768
#>  [6,] 0.26907103 0.06872791 0.26907103
#>  [7,] 0.01752469 0.01752469 0.01752469
#>  [8,] 3.34082472 0.45995322 8.09705830
#>  [9,] 0.32554947 1.83208535 0.19610180
#> [10,] 0.33252241 0.33252241 0.33252241
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "ESM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5190542 0.09501491 0.5620184
#>  [2,] 0.3814760 1.20454766 3.0895533
#>  [3,] 0.3275692 3.36888275 1.7277676
#>  [4,] 0.6319023 2.67704324 2.4253218
#>  [5,] 3.7361049 0.59100303 1.5213607
#>  [6,] 1.6947933 1.36929673 3.6837139
#>  [7,] 1.9491616 1.92097338 1.9491616
#>  [8,] 3.5954857 0.71773260 1.1367363
#>  [9,] 1.1470398 1.14703984 0.4169526
#> [10,] 0.2455533 0.26376563 0.2637656
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "LFM"
)
#>              [,1]      [,2]      [,3]
#>  [1,] 2.499399118 3.6169877 2.1072190
#>  [2,] 2.168706798 2.1687068 0.6637132
#>  [3,] 0.424566413 0.4245664 0.4245664
#>  [4,] 0.422381424 0.6538975 1.1571357
#>  [5,] 0.555550982 0.1153531 1.3654705
#>  [6,] 1.361647415 3.2294118 0.5281224
#>  [7,] 0.004229348 0.5909959 0.5909959
#>  [8,] 0.111270333 2.7602024 0.8561224
#>  [9,] 0.877275337 1.2151526 0.4773795
#> [10,] 0.364156607 1.5366971 3.5475359
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27359029 0.45275946 1.72758736
#>  [2,] 2.05820150 2.05820150 0.43894918
#>  [3,] 0.01681341 0.01681341 0.01681341
#>  [4,] 1.18172217 1.91624715 4.91947635
#>  [5,] 0.85670253 0.93927767 2.90204577
#>  [6,] 1.17849119 0.04759181 1.17849119
#>  [7,] 0.99977912 0.99977912 0.09969975
#>  [8,] 0.45843532 0.27731971 2.30828284
#>  [9,] 3.41465436 3.95951155 0.75030146
#> [10,] 3.05612110 1.74656457 0.82909268
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = c(0.5, 1e-4), family = "Pareto",
    method = "AM"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 1.6432507 0.20709483 2.66487959
#>  [2,] 3.1010247 0.33374326 0.77582309
#>  [3,] 2.1525387 0.50415665 2.54897053
#>  [4,] 2.5785608 1.46132561 2.28886560
#>  [5,] 1.0116086 0.04974407 1.90457322
#>  [6,] 0.8635255 4.06473762 0.84452088
#>  [7,] 1.7819580 1.39120543 0.06199704
#>  [8,] 2.2503659 7.37299317 1.52146918
#>  [9,] 0.8338128 0.79746618 2.00534608
#> [10,] 0.2881505 2.08883919 1.42172376

## Exponential
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential"
)
#>             [,1]        [,2]        [,3]
#>  [1,] 1.91156783 1.364972254 4.091722385
#>  [2,] 0.05842473 0.306180375 0.306180375
#>  [3,] 0.19790041 0.074659131 0.074659131
#>  [4,] 0.39943217 0.021084873 0.021084873
#>  [5,] 0.17284736 0.069269389 0.356031458
#>  [6,] 1.31800449 0.930142607 0.353142628
#>  [7,] 0.98173650 0.981736498 0.285650152
#>  [8,] 1.79975141 1.440327675 1.440327675
#>  [9,] 3.06412172 0.007590366 0.006456905
#> [10,] 0.44044410 0.440444105 0.440444105
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13620193 0.13620193 0.11671709
#>  [2,] 1.06095859 1.06095859 1.06095859
#>  [3,] 0.89615671 0.29278103 0.29278103
#>  [4,] 0.17473655 0.17473655 0.17473655
#>  [5,] 0.06015478 0.06015478 0.06015478
#>  [6,] 0.04941305 0.04941305 0.04941305
#>  [7,] 0.23526622 0.23526622 0.23526622
#>  [8,] 0.62144121 0.61923767 0.61923767
#>  [9,] 1.08110985 1.08110985 1.08110985
#> [10,] 0.19496112 0.64581301 0.64581301
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "LFM"
)
#>            [,1]      [,2]        [,3]
#>  [1,] 0.3920371 1.6756126 0.001248105
#>  [2,] 0.5958927 0.5958927 0.595892676
#>  [3,] 1.1742475 0.3212596 0.321259606
#>  [4,] 0.2551419 2.1379690 0.255141943
#>  [5,] 0.4077472 0.4077472 0.675785025
#>  [6,] 0.2445223 0.2445223 0.244522258
#>  [7,] 0.2800700 0.2800700 0.018276384
#>  [8,] 0.2322473 0.4116558 0.516009443
#>  [9,] 0.3531967 0.3531967 0.353196660
#> [10,] 0.0382379 0.1101226 0.437472919
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.12612873 0.01403341 1.12612873
#>  [2,] 1.09437447 0.96122787 1.09437447
#>  [3,] 0.05942810 0.17711557 0.17711557
#>  [4,] 0.27494034 0.27494034 0.19292631
#>  [5,] 0.57490032 0.57490032 0.57490032
#>  [6,] 0.05103869 0.05103869 0.05103869
#>  [7,] 1.43732342 1.37104903 1.43732342
#>  [8,] 0.28944985 0.47611755 1.65241262
#>  [9,] 0.32701941 0.32701941 0.32701941
#> [10,] 0.94673133 2.37384542 1.30493590
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Exponential",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03032097 0.03032097 0.03032097
#>  [2,] 0.55920570 0.50041860 0.50041860
#>  [3,] 0.98527002 0.98527002 0.41911028
#>  [4,] 0.32581965 0.32581965 0.32581965
#>  [5,] 0.29839737 0.10321021 0.29839737
#>  [6,] 0.15084936 0.15084936 0.15084936
#>  [7,] 0.63366044 1.12979989 1.12979989
#>  [8,] 0.04602624 0.26409291 0.02251379
#>  [9,] 0.06798271 0.06798271 0.06798271
#> [10,] 0.58794456 0.66673407 0.66673407

## Alpha-Stable
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 0.265792619 0.265792619 1.351581422
#>  [2,] 1.348045675 0.418651724 0.136619968
#>  [3,] 1.188650862 0.036727479 0.876816345
#>  [4,] 0.019912010 0.019912010 0.019912010
#>  [5,] 0.803139033 0.407101204 0.422796778
#>  [6,] 0.009088993 0.009088993 0.009088993
#>  [7,] 0.530083654 0.072176599 0.030079779
#>  [8,] 0.332049678 0.332049678 0.332049678
#>  [9,] 0.171034887 0.171034887 0.171034887
#> [10,] 0.436381159 0.836601841 0.836601841
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27475862 0.27475862 1.34071816
#>  [2,] 0.24535719 0.24535719 0.24535719
#>  [3,] 0.28932401 0.90524130 1.25069713
#>  [4,] 0.06584131 0.06584131 0.06584131
#>  [5,] 0.06595952 0.22165691 0.13452045
#>  [6,] 0.53144004 0.28723000 0.28723000
#>  [7,] 0.95056217 0.08387134 0.12607452
#>  [8,] 1.68628715 1.64324369 1.68628715
#>  [9,] 0.40354529 0.37855363 0.40354529
#> [10,] 0.66272980 0.08894683 0.34627086
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06937172 0.28126061 0.17934031
#>  [2,] 0.35796759 0.35796759 0.25837202
#>  [3,] 0.04133449 0.44686638 0.49236698
#>  [4,] 1.03838297 0.56255924 1.03838297
#>  [5,] 0.29304305 0.29304305 1.35322299
#>  [6,] 0.44009733 0.67707128 1.38150360
#>  [7,] 0.86038491 1.85922585 0.64185890
#>  [8,] 0.41839629 0.31359310 0.15093032
#>  [9,] 0.39803963 2.09382218 0.48991762
#> [10,] 0.01842769 0.03953839 0.01842769
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "AlphaStable",
    method = "AM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.24039881 1.09795491 0.126985363
#>  [2,] 0.69533697 0.28479205 0.002850727
#>  [3,] 0.01187807 0.01187807 0.765389873
#>  [4,] 0.60869318 0.04652949 0.121720840
#>  [5,] 0.08460993 0.34894994 0.012036672
#>  [6,] 0.15587809 0.27679413 0.276794128
#>  [7,] 0.31282894 0.51349479 0.513494793
#>  [8,] 0.54773058 0.54773058 0.040756893
#>  [9,] 0.27878725 0.27878725 0.278787247
#> [10,] 0.16375497 0.92602438 0.388403920

## Inverse Gaussian
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.25813708 0.25813708 0.02149860
#>  [2,] 0.79072409 0.51592255 0.94843486
#>  [3,] 0.93074986 0.88758797 1.00348203
#>  [4,] 1.29718977 0.28614303 1.29718977
#>  [5,] 0.46774500 0.02499206 0.36446936
#>  [6,] 0.12873068 0.12873068 0.12873068
#>  [7,] 0.50913592 0.50913592 0.50913592
#>  [8,] 0.57634547 0.03707534 0.57634547
#>  [9,] 0.60017353 0.45631576 1.68605699
#> [10,] 0.07747135 0.03897418 0.07747135
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.08008936 0.90855706 0.18450218
#>  [2,] 0.51090973 0.51090973 0.44899186
#>  [3,] 0.29034832 0.36979601 0.36979601
#>  [4,] 1.22683631 0.35409125 0.35409125
#>  [5,] 0.49045143 0.11283256 0.42967079
#>  [6,] 0.12052690 0.01894284 0.12052690
#>  [7,] 0.44701027 0.58048620 0.40665384
#>  [8,] 0.41492441 0.64225961 0.17344886
#>  [9,] 0.03217132 0.03217132 0.03217132
#> [10,] 0.11702252 0.11702252 0.11483417
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.26919869 0.32071101 0.3154921
#>  [2,] 0.25146080 0.22545580 0.2943336
#>  [3,] 0.13151968 0.33284309 0.3328431
#>  [4,] 0.27438929 0.26473898 0.3620829
#>  [5,] 0.56997135 0.51053000 0.9301541
#>  [6,] 0.21329309 0.21329309 0.1457580
#>  [7,] 1.16907722 0.02389626 0.1431273
#>  [8,] 0.04664882 0.04664882 1.0532560
#>  [9,] 0.65542994 0.66317117 0.6554299
#> [10,] 0.53123469 0.17020274 0.5312347
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "InverseGaussian",
    method = "AM"
)
#>              [,1]       [,2]       [,3]
#>  [1,] 0.829619531 0.07743634 0.62166882
#>  [2,] 0.400503809 0.15750826 0.40050381
#>  [3,] 1.856589924 0.59019045 0.13573879
#>  [4,] 0.408242346 0.40824235 0.01190570
#>  [5,] 0.680509067 0.69341307 1.00149674
#>  [6,] 0.344762580 0.10588599 0.34476258
#>  [7,] 0.002180543 0.06777949 0.26972662
#>  [8,] 0.503118320 0.50311832 0.50311832
#>  [9,] 0.096171862 0.45895202 0.03149253
#> [10,] 0.490215547 0.20089518 0.47128181

## Gamma
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03039095 0.03039095 0.34837888
#>  [2,] 0.56353549 0.67199317 0.12472683
#>  [3,] 0.20089575 0.02812678 0.02812678
#>  [4,] 0.39817247 0.48688907 0.39817247
#>  [5,] 1.21890663 0.50260294 0.50260294
#>  [6,] 0.43158634 0.77045785 0.10162315
#>  [7,] 0.53236423 0.53236423 0.53236423
#>  [8,] 0.21397726 0.24509917 0.24509917
#>  [9,] 0.15685562 0.04507333 1.23862867
#> [10,] 0.25834928 0.55049357 0.25834928
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.62169291 0.9503756 0.3635764
#>  [2,] 0.34897275 0.4678790 0.4678790
#>  [3,] 0.69270892 0.7664173 0.3607099
#>  [4,] 0.22611400 0.3490195 0.3490195
#>  [5,] 0.36057846 0.6965301 1.2177893
#>  [6,] 0.02044016 0.5505587 0.5505587
#>  [7,] 0.39327520 0.4027406 0.7140072
#>  [8,] 0.03434657 0.1131959 0.1131959
#>  [9,] 0.82492716 0.6985204 0.5552797
#> [10,] 0.30650569 0.3065057 0.3065057
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "MDCM"
)
#>             [,1]        [,2]        [,3]
#>  [1,] 0.21719436 0.007726223 0.007726223
#>  [2,] 0.04012457 0.040124568 0.040124568
#>  [3,] 0.08433673 0.275641180 2.234965413
#>  [4,] 0.68401182 0.908069633 0.542326118
#>  [5,] 0.19481647 0.194816469 0.173135540
#>  [6,] 0.06970305 0.069703052 0.301907371
#>  [7,] 0.11613911 0.024295385 0.226144244
#>  [8,] 0.31177293 1.124362830 0.359679946
#>  [9,] 0.60183873 0.601838732 0.601838732
#> [10,] 0.14162478 0.141624778 0.061764479
rpextmo(
    10, 3,
    a = 0.2, b = 0.5, gamma = 2,
    eta = 0.5, family = "Gamma",
    method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.03144116 0.05566957 0.05566957
#>  [2,] 0.03263889 0.14236387 0.85689295
#>  [3,] 0.07855835 0.85506522 0.11675326
#>  [4,] 0.27418868 0.27418868 0.31139605
#>  [5,] 0.05698188 0.05698188 0.01990472
#>  [6,] 0.38907933 0.38907933 0.38907933
#>  [7,] 0.02716187 0.72248584 0.26395161
#>  [8,] 0.11476529 0.11476529 0.03329045
#>  [9,] 0.60397461 0.60397461 0.60397461
#> [10,] 0.18150352 0.58383504 0.79250095