Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3050876 1.2925452 1.0001356
#>  [2,] 0.3190545 0.7792098 3.9040841
#>  [3,] 2.2270788 2.5148611 2.5148611
#>  [4,] 0.6261240 1.0374786 1.0374786
#>  [5,] 3.0238064 0.3333150 0.4078451
#>  [6,] 0.3661040 0.3661040 0.3661040
#>  [7,] 1.7611803 0.9624885 2.2374364
#>  [8,] 0.6476698 1.6169400 1.2703381
#>  [9,] 4.3204034 2.2736934 3.5419594
#> [10,] 0.9145148 4.5981315 4.5981315
rpextmo(10, 3, a = 1)      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3189881 0.3189881 0.3189881
#>  [2,] 2.5147184 2.5147184 2.5147184
#>  [3,] 1.0401954 1.0401954 1.0401954
#>  [4,] 0.2192257 0.2192257 0.2192257
#>  [5,] 0.6191052 0.6191052 0.6191052
#>  [6,] 0.9311572 0.9311572 0.9311572
#>  [7,] 0.4276145 0.4276145 0.4276145
#>  [8,] 0.9771905 0.9771905 0.9771905
#>  [9,] 1.3784685 1.3784685 1.3784685
#> [10,] 0.6699834 0.6699834 0.6699834
rpextmo(10, 3, b = 1)      ## independence
#>              [,1]        [,2]      [,3]
#>  [1,] 0.330490956 1.593820462 0.1012852
#>  [2,] 0.173675203 0.738168589 1.5503970
#>  [3,] 4.112995377 0.066752906 0.8437952
#>  [4,] 1.095357549 1.531169433 0.2546377
#>  [5,] 0.630784511 0.023776079 0.2896385
#>  [6,] 0.724424729 0.820391138 0.5991645
#>  [7,] 0.513088769 1.135739231 0.1976715
#>  [8,] 0.008762037 1.331741762 0.3537982
#>  [9,] 0.665929473 1.593718195 0.5663940
#> [10,] 1.101344843 0.009893157 0.6911094

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.3467144 0.68368469 2.3238545
#>  [2,] 1.2297724 1.22977236 1.2297724
#>  [3,] 0.9630215 0.22122138 1.5430446
#>  [4,] 0.2097410 1.01108124 1.8292174
#>  [5,] 0.4749655 3.81635390 3.7441039
#>  [6,] 0.5389734 0.06412818 1.1287033
#>  [7,] 1.8097487 1.42342257 1.2096290
#>  [8,] 0.8209232 0.35191460 0.2426977
#>  [9,] 1.6100969 0.36870398 1.6100969
#> [10,] 0.8679482 2.26431044 1.1917566
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.6055625 0.6055625 0.6055625
#>  [2,] 0.1559455 0.1559455 0.1559455
#>  [3,] 0.4317439 0.4317439 0.4317439
#>  [4,] 0.8501113 0.8501113 0.8501113
#>  [5,] 0.3155636 0.3155636 0.3155636
#>  [6,] 0.7726141 0.7726141 0.7726141
#>  [7,] 1.2242668 1.2242668 1.2242668
#>  [8,] 1.9203640 1.9203640 1.9203640
#>  [9,] 4.1061642 4.1061642 4.1061642
#> [10,] 1.3345605 1.3345605 1.3345605
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.95774267 2.1402055 0.6767005
#>  [2,] 0.42028699 0.8333268 1.4210837
#>  [3,] 0.65629374 0.5303704 1.6930955
#>  [4,] 1.43717103 2.4813685 0.4708418
#>  [5,] 0.20989134 2.1425753 0.6898324
#>  [6,] 0.02298275 0.8182257 0.2860650
#>  [7,] 1.10378483 1.8143217 0.1291433
#>  [8,] 1.58490817 0.3016274 0.3601668
#>  [9,] 0.48554304 8.9760532 0.2994177
#> [10,] 1.10397843 1.1635881 0.3761604

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3284759 5.9604523 5.99358084
#>  [2,] 0.4078420 1.9391277 3.44589627
#>  [3,] 5.2011666 1.4642055 0.68533142
#>  [4,] 0.9320619 1.7246359 0.69664983
#>  [5,] 4.0141621 4.0141621 0.69044815
#>  [6,] 1.0712080 0.8660625 0.07953622
#>  [7,] 0.8881379 0.8881379 0.88813793
#>  [8,] 6.2322259 2.5558594 0.15741964
#>  [9,] 1.4444851 3.2473969 1.56277157
#> [10,] 2.7134417 2.9521632 0.75873188
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.39161625 1.39161625 1.39161625
#>  [2,] 2.44905111 2.44905111 2.44905111
#>  [3,] 2.08583901 2.08583901 2.08583901
#>  [4,] 0.04139050 0.04139050 0.04139050
#>  [5,] 0.01127179 0.01127179 0.01127179
#>  [6,] 0.77520869 0.77520869 0.77520869
#>  [7,] 1.20928018 1.20928018 1.20928018
#>  [8,] 2.27472579 2.27472579 2.27472579
#>  [9,] 1.54205268 1.54205268 1.54205268
#> [10,] 3.32590026 3.32590026 3.32590026
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]        [,2]      [,3]
#>  [1,] 1.46532588 1.009930072 1.0588147
#>  [2,] 6.19451140 1.475011678 0.8654094
#>  [3,] 0.29647423 0.007875542 1.9474343
#>  [4,] 0.06340421 1.301432156 0.7929885
#>  [5,] 2.93166874 0.173247812 0.1020009
#>  [6,] 0.04560089 0.815489530 0.2215646
#>  [7,] 3.34963620 0.134689315 0.2970243
#>  [8,] 0.53454226 0.015416439 0.8130466
#>  [9,] 1.11965974 1.355106814 0.3163160
#> [10,] 1.39486264 0.661977856 2.1591636

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.9042439 0.02487620 0.93521326
#>  [2,] 2.0673995 0.20784708 1.03413116
#>  [3,] 3.8045458 6.36117912 4.54445592
#>  [4,] 1.6728530 0.09591853 0.07479977
#>  [5,] 0.1920441 0.36021084 0.06103031
#>  [6,] 2.1783196 0.52637967 0.33408613
#>  [7,] 1.9717615 0.23505942 0.46099321
#>  [8,] 0.7160826 2.67000272 3.02211619
#>  [9,] 2.6531954 1.13172674 0.33414721
#> [10,] 0.4360434 1.03498302 0.31292425
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58916283 0.58916283 0.58916283
#>  [2,] 0.44364111 0.44364111 0.44364111
#>  [3,] 2.23561913 2.23561913 2.23561913
#>  [4,] 0.01735721 0.01735721 0.01735721
#>  [5,] 0.24137393 0.24137393 0.24137393
#>  [6,] 0.12154242 0.12154242 0.12154242
#>  [7,] 0.87747592 0.87747592 0.87747592
#>  [8,] 0.56042780 0.56042780 0.56042780
#>  [9,] 0.30719800 0.30719800 0.30719800
#> [10,] 1.01963905 1.01963905 1.01963905
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.40482160 1.90655419 0.97352739
#>  [2,] 1.21196348 0.26580777 0.22298919
#>  [3,] 0.21777880 0.80026770 0.16826493
#>  [4,] 0.97065190 0.20369140 0.09402854
#>  [5,] 0.47433948 0.57759126 0.29710282
#>  [6,] 3.78703073 0.66360170 0.08602558
#>  [7,] 0.03234045 0.37079480 1.03843125
#>  [8,] 4.57756646 0.38963629 0.75865738
#>  [9,] 0.41409178 0.80863061 0.81239282
#> [10,] 0.10997029 0.07968029 1.93977750

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.821984612 1.1491779 1.1491779
#>  [2,] 2.569934416 1.3243828 3.2875117
#>  [3,] 4.071174304 2.7969677 3.3095710
#>  [4,] 1.784891710 0.0815238 0.1428933
#>  [5,] 1.161441016 1.1614410 0.5044402
#>  [6,] 0.006530009 4.0452054 2.0233724
#>  [7,] 0.219961351 1.6076138 0.4028279
#>  [8,] 0.803733781 0.9091110 0.6858444
#>  [9,] 4.424289951 0.4501972 2.6225115
#> [10,] 0.321605547 4.0985143 0.1822935
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.14889644 2.14889644 2.14889644
#>  [2,] 0.02709055 0.02709055 0.02709055
#>  [3,] 3.24718763 3.24718763 3.24718763
#>  [4,] 0.48525369 0.48525369 0.48525369
#>  [5,] 0.86710805 0.86710805 0.86710805
#>  [6,] 0.30002690 0.30002690 0.30002690
#>  [7,] 0.20662791 0.20662791 0.20662791
#>  [8,] 3.01680507 3.01680507 3.01680507
#>  [9,] 0.50772808 0.50772808 0.50772808
#> [10,] 0.35166373 0.35166373 0.35166373
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 1.69538013 2.6474844 1.70030578
#>  [2,] 2.44157263 1.0040261 0.14678414
#>  [3,] 0.66825050 0.4779350 0.98120616
#>  [4,] 1.14699482 0.2245046 0.51340037
#>  [5,] 0.30205327 0.4849028 2.70308295
#>  [6,] 3.97833119 0.6924486 0.57957372
#>  [7,] 0.27592886 0.3295067 0.60623682
#>  [8,] 1.07092168 1.6703203 0.03165756
#>  [9,] 1.41517102 0.4158956 1.06691905
#> [10,] 0.03561095 1.1640626 1.58104521

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]      [,2]      [,3]
#>  [1,] 5.5524741 6.1530095 1.5811410
#>  [2,] 2.1599023 0.7668549 2.2159597
#>  [3,] 1.0556572 0.5708137 1.2041365
#>  [4,] 2.0085461 3.9393364 1.2282193
#>  [5,] 1.2983688 3.0834979 1.0975276
#>  [6,] 1.1646417 0.1447431 0.5493632
#>  [7,] 1.1813795 2.7295824 0.6429141
#>  [8,] 0.1110991 1.4194345 0.7069252
#>  [9,] 0.3978292 1.6595712 0.5261661
#> [10,] 0.3133322 1.7084456 0.2888488
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.99775710 0.99775710 0.99775710
#>  [2,] 0.36833123 0.36833123 0.36833123
#>  [3,] 0.29719775 0.29719775 0.29719775
#>  [4,] 0.04371778 0.04371778 0.04371778
#>  [5,] 0.33556163 0.33556163 0.33556163
#>  [6,] 0.68717887 0.68717887 0.68717887
#>  [7,] 0.35733539 0.35733539 0.35733539
#>  [8,] 0.76724798 0.76724798 0.76724798
#>  [9,] 0.37464077 0.37464077 0.37464077
#> [10,] 0.54767361 0.54767361 0.54767361
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.06267046 0.1808666 2.2388863
#>  [2,] 0.19311792 0.2197450 1.1533412
#>  [3,] 0.33238220 1.8298280 1.1216194
#>  [4,] 0.05658093 0.0309108 0.5675369
#>  [5,] 0.52823232 0.3925668 0.3153467
#>  [6,] 0.69743754 0.2316511 0.2296477
#>  [7,] 0.41678712 1.5684569 0.6884423
#>  [8,] 1.18622151 0.6181724 0.7932820
#>  [9,] 0.27594982 3.1361438 0.1843074
#> [10,] 1.01755292 3.0882851 0.8951864

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]         [,2]      [,3]
#>  [1,] 2.1617573 0.1358489831 0.6657923
#>  [2,] 0.1078887 3.5132252647 2.4639332
#>  [3,] 1.1582856 0.4169869088 0.9337017
#>  [4,] 3.4900399 3.4900399249 0.3779139
#>  [5,] 0.6290751 0.0764910871 0.5622476
#>  [6,] 0.7954333 2.7440633858 2.7440634
#>  [7,] 2.3327721 1.4143955241 1.5011362
#>  [8,] 0.1420890 2.0309416241 1.3159004
#>  [9,] 4.7557538 3.6624875365 2.1330319
#> [10,] 1.1172115 0.0008034268 6.3908132
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05269619 0.05269619 0.05269619
#>  [2,] 4.11330303 4.11330303 4.11330303
#>  [3,] 0.05389383 0.05389383 0.05389383
#>  [4,] 0.64497927 0.64497927 0.64497927
#>  [5,] 0.47658581 0.47658581 0.47658581
#>  [6,] 1.53155641 1.53155641 1.53155641
#>  [7,] 3.19480955 3.19480955 3.19480955
#>  [8,] 1.95307608 1.95307608 1.95307608
#>  [9,] 0.23437583 0.23437583 0.23437583
#> [10,] 2.58581323 2.58581323 2.58581323
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.9055308 0.58820206 1.05103662
#>  [2,] 0.4298642 3.43375159 0.57535947
#>  [3,] 2.2936590 1.69071710 1.58941321
#>  [4,] 0.5517546 0.18125966 0.23731568
#>  [5,] 1.3635061 2.33799668 0.08560051
#>  [6,] 0.1977248 0.28398814 0.13398623
#>  [7,] 0.5220447 3.51814789 2.73959292
#>  [8,] 4.0025532 0.41092216 0.79720365
#>  [9,] 3.8227048 0.49765377 2.04060200
#> [10,] 2.8356014 0.02628569 0.96417908

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4907320 1.35226936 0.8032635
#>  [2,] 0.1537389 0.33753755 0.3375375
#>  [3,] 1.5980392 0.39722425 0.5591044
#>  [4,] 2.3862671 5.40815407 2.7224772
#>  [5,] 0.2780778 6.35274349 0.1699576
#>  [6,] 6.7593173 0.41160956 0.5000519
#>  [7,] 0.7478330 2.93522805 1.8339632
#>  [8,] 1.8985399 0.08151933 1.7109827
#>  [9,] 0.5001389 0.55228624 0.5522862
#> [10,] 2.0130739 0.53484824 2.6202339
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.2325355 1.2325355 1.2325355
#>  [2,] 5.2709279 5.2709279 5.2709279
#>  [3,] 2.7214297 2.7214297 2.7214297
#>  [4,] 0.5319481 0.5319481 0.5319481
#>  [5,] 1.2060454 1.2060454 1.2060454
#>  [6,] 0.9172913 0.9172913 0.9172913
#>  [7,] 0.5462169 0.5462169 0.5462169
#>  [8,] 0.6642028 0.6642028 0.6642028
#>  [9,] 1.6142074 1.6142074 1.6142074
#> [10,] 1.3372227 1.3372227 1.3372227
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>            [,1]        [,2]      [,3]
#>  [1,] 1.0521733 0.188451132 0.5446364
#>  [2,] 0.1032521 0.163609825 0.5592820
#>  [3,] 1.0783993 0.074992006 2.7345259
#>  [4,] 0.6109967 0.181681534 2.7429970
#>  [5,] 1.9976404 4.604915905 0.3881974
#>  [6,] 0.4016517 0.205477856 0.6518907
#>  [7,] 4.8801037 0.925923096 2.1588287
#>  [8,] 3.1965095 0.083891207 0.3111900
#>  [9,] 3.7486087 0.716051932 0.9307296
#> [10,] 0.0948446 0.002052424 0.9627930

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.73164418 0.7328050 0.73280502
#>  [2,] 0.06787528 4.0113985 1.04610481
#>  [3,] 2.36300817 1.4359108 1.84150505
#>  [4,] 1.88294969 3.2253018 0.08389756
#>  [5,] 0.57547891 1.5546195 3.79551546
#>  [6,] 0.86074825 1.6952233 1.57775057
#>  [7,] 0.65068342 1.2742132 0.39115168
#>  [8,] 1.75507790 3.4212411 0.61718199
#>  [9,] 0.05876770 0.7930375 0.79303748
#> [10,] 1.66120729 1.6062692 0.54850741
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47706384 0.47706384 0.47706384
#>  [2,] 0.34999934 0.34999934 0.34999934
#>  [3,] 1.54079289 1.54079289 1.54079289
#>  [4,] 2.01325219 2.01325219 2.01325219
#>  [5,] 0.10396744 0.10396744 0.10396744
#>  [6,] 0.19019715 0.19019715 0.19019715
#>  [7,] 1.70721782 1.70721782 1.70721782
#>  [8,] 0.21841176 0.21841176 0.21841176
#>  [9,] 0.03869927 0.03869927 0.03869927
#> [10,] 0.77546187 0.77546187 0.77546187
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]        [,2]       [,3]
#>  [1,] 1.5506312 2.082741702 0.94216995
#>  [2,] 0.2812879 0.487164736 1.05802324
#>  [3,] 0.9520414 0.137440721 0.22384264
#>  [4,] 0.3519673 0.136716532 0.04148088
#>  [5,] 0.5190103 0.116761342 0.78331584
#>  [6,] 1.0009890 0.009238417 1.32095563
#>  [7,] 1.6903520 0.676093881 0.51201472
#>  [8,] 0.5573879 0.183266435 0.13501989
#>  [9,] 2.2849799 0.011621389 1.17617708
#> [10,] 0.4262229 0.908542590 0.46341488

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>             [,1]       [,2]         [,3]
#>  [1,] 1.84964512 0.73677683 0.4241223567
#>  [2,] 1.28951183 1.11024686 1.7461794392
#>  [3,] 0.41048221 1.57153163 4.1570969574
#>  [4,] 5.83886841 1.34592696 1.3144120933
#>  [5,] 1.00982014 1.00982014 0.3876938512
#>  [6,] 0.08416462 1.31538488 2.4425872228
#>  [7,] 0.03292170 2.67362925 1.7180218451
#>  [8,] 2.59284142 0.01665372 1.7129734032
#>  [9,] 2.74941012 0.24881845 0.0004214769
#> [10,] 0.65471654 0.65471654 0.6547165381
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9704559 0.9704559 0.9704559
#>  [2,] 1.1231757 1.1231757 1.1231757
#>  [3,] 1.6000665 1.6000665 1.6000665
#>  [4,] 3.9688120 3.9688120 3.9688120
#>  [5,] 1.8932127 1.8932127 1.8932127
#>  [6,] 1.2784864 1.2784864 1.2784864
#>  [7,] 5.0776193 5.0776193 5.0776193
#>  [8,] 0.4824571 0.4824571 0.4824571
#>  [9,] 0.3184523 0.3184523 0.3184523
#> [10,] 1.5680367 1.5680367 1.5680367
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]      [,2]      [,3]
#>  [1,] 1.1372842 1.7141304 2.0842743
#>  [2,] 0.2660934 0.8159038 2.1988365
#>  [3,] 1.1984083 0.3243484 0.3496583
#>  [4,] 0.2552877 1.8246083 0.9959498
#>  [5,] 0.9175970 4.8736290 3.1339877
#>  [6,] 1.1941879 0.4133670 0.8676454
#>  [7,] 1.2008098 0.7392486 0.1772546
#>  [8,] 4.0845264 0.6879673 0.4431001
#>  [9,] 1.7743315 0.8782512 0.4622661
#> [10,] 0.3487967 2.0699894 1.9446172

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5297732 1.00121755 0.95862513
#>  [2,] 0.6856899 0.71725303 1.86535758
#>  [3,] 0.8876015 0.37666682 0.52651947
#>  [4,] 0.3993395 0.47464261 0.39933947
#>  [5,] 0.8362127 0.32523258 0.32523258
#>  [6,] 0.9381693 0.02592891 0.02592891
#>  [7,] 0.7191635 0.71916346 0.04087377
#>  [8,] 2.0195270 0.07207735 0.09142061
#>  [9,] 4.1240502 1.37591988 0.04240469
#> [10,] 0.2424233 0.16931224 1.23453495
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.07194081 0.959007350 0.071940809
#>  [2,] 0.29387077 0.253670511 2.248942437
#>  [3,] 0.47863345 0.950626996 0.458232270
#>  [4,] 0.18146007 0.786955406 0.573716443
#>  [5,] 0.53558419 0.613670915 0.105990896
#>  [6,] 1.60070814 0.983406517 0.004927984
#>  [7,] 0.91934347 0.684849230 0.118697702
#>  [8,] 0.42649981 0.371843180 0.868178318
#>  [9,] 1.92183630 0.009082441 0.655685271
#> [10,] 0.97364530 0.973645301 2.883398298
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.65035976 0.21220887 3.38111342
#>  [2,] 0.33758893 1.34411223 0.33758893
#>  [3,] 1.56703944 2.20685543 0.14690223
#>  [4,] 0.64977738 1.80931127 2.06460307
#>  [5,] 0.03988953 0.71252921 0.96977765
#>  [6,] 0.07772268 0.07772268 0.07772268
#>  [7,] 1.08044403 0.24285700 0.57283214
#>  [8,] 0.53525178 0.29441911 0.22476916
#>  [9,] 0.74585934 0.74585934 0.62594186
#> [10,] 0.50229262 0.16672247 0.18190676
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.34264261 0.53302627 0.86621641
#>  [2,] 0.02465101 0.88072855 0.88072855
#>  [3,] 0.34483418 0.49205785 0.01199374
#>  [4,] 0.01511139 0.01511139 0.01511139
#>  [5,] 0.62620163 0.52058561 0.53995772
#>  [6,] 0.97086584 1.04142480 1.28407084
#>  [7,] 0.61499949 0.61499949 0.74034163
#>  [8,] 0.17450154 0.97453973 0.40356654
#>  [9,] 0.15524952 0.91206122 1.50173462
#> [10,] 0.44630125 0.49233265 0.44630125
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.9431248 0.94312480 0.7363485
#>  [2,] 0.2931136 0.32741483 0.3346235
#>  [3,] 0.5440298 0.03811302 0.5440298
#>  [4,] 0.2519636 0.25196357 0.2519636
#>  [5,] 0.1731689 0.11591145 0.1731689
#>  [6,] 0.3824644 0.73480738 0.3674285
#>  [7,] 0.8315985 0.75148978 0.8315985
#>  [8,] 0.0436182 0.02739848 0.0436182
#>  [9,] 0.4057161 0.40571607 0.4122051
#> [10,] 1.2285189 1.78300879 1.1445806

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.1979048 1.1979048 0.21404662
#>  [2,] 2.0245572 1.1332784 1.81459860
#>  [3,] 0.9297121 1.8804384 0.05802902
#>  [4,] 2.4731754 0.5442694 0.25550197
#>  [5,] 1.2130034 0.1067350 1.21300337
#>  [6,] 2.2652057 6.1096055 0.20158457
#>  [7,] 3.4486340 0.9480571 0.46505733
#>  [8,] 0.0672815 0.5076218 0.03815027
#>  [9,] 1.2065241 0.1742462 6.59772667
#> [10,] 0.4987938 0.4987938 0.49879378
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.2203220 0.2203220 0.22032195
#>  [2,] 0.1540651 1.0185491 0.64986344
#>  [3,] 2.1370095 0.2161655 1.94956507
#>  [4,] 1.7908518 1.6338193 1.20888610
#>  [5,] 1.3014141 1.2276287 1.35217169
#>  [6,] 4.3274544 0.1136576 1.21377187
#>  [7,] 2.0559211 0.2657556 0.18460129
#>  [8,] 0.9461945 2.0936225 0.35652870
#>  [9,] 2.5735419 2.1603853 4.53073396
#> [10,] 0.3889574 0.3429786 0.03565769
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.55141572 2.55141572 2.55141572
#>  [2,] 1.43379184 0.51725468 1.43379184
#>  [3,] 0.71764000 4.15665207 1.93813538
#>  [4,] 2.38493098 1.83392416 1.55780072
#>  [5,] 0.03784104 0.02374858 0.03784104
#>  [6,] 0.46179537 0.03364475 1.61578587
#>  [7,] 0.58545865 1.36283592 1.36283592
#>  [8,] 2.95776505 1.85094215 0.30751859
#>  [9,] 0.25319958 2.90864776 0.30870498
#> [10,] 0.64496910 0.28639044 0.64496910
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.9571769 0.82567613 0.7169930
#>  [2,] 0.2013294 4.37011603 1.5659817
#>  [3,] 5.9259472 1.64788031 2.0932797
#>  [4,] 2.8014047 1.54768934 0.2485561
#>  [5,] 0.8508125 0.20365408 0.2800700
#>  [6,] 0.4357877 2.93432497 1.3515179
#>  [7,] 4.7446414 1.21642762 3.4639010
#>  [8,] 0.7613694 0.51339122 3.2891625
#>  [9,] 0.2980902 2.09291384 1.4475970
#> [10,] 2.5297785 0.08410013 1.7416384
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>             [,1]      [,2]      [,3]
#>  [1,] 1.03943054 3.4476336 4.7589621
#>  [2,] 3.57367483 0.5283107 3.6966899
#>  [3,] 0.86573680 1.2295920 2.2359526
#>  [4,] 0.19998591 0.9501556 1.6069760
#>  [5,] 0.30643828 0.3064383 0.2810853
#>  [6,] 0.42321700 0.4232170 0.3765062
#>  [7,] 2.71118390 0.3631989 1.5985806
#>  [8,] 5.61425820 2.6042134 0.2387861
#>  [9,] 2.05157842 0.8920839 0.1497336
#> [10,] 0.04908198 1.1734027 1.1289161

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.66442161 0.61659891 1.59174565
#>  [2,] 0.28126080 0.11082082 0.81584405
#>  [3,] 0.91188505 2.06127242 2.06127242
#>  [4,] 0.35722273 0.35722273 0.76372416
#>  [5,] 0.37826803 0.65599167 0.65599167
#>  [6,] 1.31284337 1.31284337 0.02309683
#>  [7,] 0.04856275 0.04856275 0.04856275
#>  [8,] 0.30920281 0.16987474 0.82423407
#>  [9,] 0.24342645 0.24342645 0.24342645
#> [10,] 0.74894229 0.74894229 0.51041170
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0139175 0.4339802 0.3079310
#>  [2,] 0.4909703 1.4798254 0.4909703
#>  [3,] 0.2630328 0.2630328 0.2630328
#>  [4,] 0.3114341 0.3114341 0.3040632
#>  [5,] 0.4292190 0.1810473 0.8806081
#>  [6,] 0.3661281 0.3661281 0.3661281
#>  [7,] 0.2660375 1.4461987 1.4461987
#>  [8,] 0.6159129 1.5433142 0.2002836
#>  [9,] 0.4012947 0.4034485 0.4012947
#> [10,] 0.4131538 0.4131538 0.2515978
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>            [,1]        [,2]       [,3]
#>  [1,] 1.4213987 0.055945879 1.02312047
#>  [2,] 0.2674027 0.267402657 0.26740266
#>  [3,] 0.2579350 0.009913187 0.25793496
#>  [4,] 0.1858839 0.185883930 0.09953932
#>  [5,] 0.6384813 0.243646923 1.46600336
#>  [6,] 0.1613382 0.161338166 0.16133817
#>  [7,] 0.1007256 0.100725553 0.10072555
#>  [8,] 0.1671469 0.167146926 0.16714693
#>  [9,] 0.4583788 0.954994984 1.37393943
#> [10,] 0.2872593 0.257092206 0.28725927
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.13610365 0.35916730 0.35916730
#>  [2,] 0.50516392 0.81557674 0.50516392
#>  [3,] 0.27722769 0.85098657 0.85098657
#>  [4,] 0.23769730 0.23769730 0.23769730
#>  [5,] 1.05854569 1.05854569 1.05854569
#>  [6,] 0.65193057 1.20239607 0.47672420
#>  [7,] 0.88457595 0.88457595 0.88457595
#>  [8,] 0.75872431 0.20695242 0.20695242
#>  [9,] 0.05346673 0.05346673 0.05346673
#> [10,] 0.16375409 0.16375409 0.16375409
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.011154221 0.01115422 0.011154221
#>  [2,] 0.065896919 0.06589692 0.177012404
#>  [3,] 0.006806173 0.58433378 0.006806173
#>  [4,] 2.531396483 0.91171964 1.776195955
#>  [5,] 0.157482061 0.39656562 0.880822786
#>  [6,] 0.913434041 0.91343404 0.913434041
#>  [7,] 0.133561989 0.13356199 0.133561989
#>  [8,] 1.339630858 0.29509367 1.339630858
#>  [9,] 0.470934956 1.57940163 1.579401631
#> [10,] 1.215654392 1.55676501 1.215654392

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14526332 0.14526332 0.14526332
#>  [2,] 0.11709172 0.11709172 0.11709172
#>  [3,] 0.23626416 0.20433474 0.23626416
#>  [4,] 0.31798746 0.80321848 1.22532258
#>  [5,] 0.21141725 0.21141725 0.21141725
#>  [6,] 0.10969965 0.10969965 0.10969965
#>  [7,] 0.26042609 0.26042609 0.18024286
#>  [8,] 0.33591910 0.71666024 0.71329868
#>  [9,] 0.83862376 0.31320423 0.83862376
#> [10,] 0.04170164 0.04170164 0.04170164
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.15547303 0.1815963 0.45086211
#>  [2,] 0.30572910 0.3363269 0.42043427
#>  [3,] 0.09861613 0.1652516 0.09789070
#>  [4,] 0.46154439 0.4493554 0.46154439
#>  [5,] 0.19149860 0.3879396 0.02285215
#>  [6,] 0.12835064 0.1283506 0.12835064
#>  [7,] 0.13030246 0.2053857 0.24676080
#>  [8,] 0.39483635 0.1167504 0.39483635
#>  [9,] 0.22162776 0.2106107 0.21061075
#> [10,] 0.57685539 0.5768554 0.33032499
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.288448259 0.288448259 0.049744698
#>  [2,] 0.056202967 0.056202967 0.056202967
#>  [3,] 0.521758166 0.521758166 0.368379332
#>  [4,] 0.326560067 0.009249436 0.326560067
#>  [5,] 0.008801571 0.008801571 0.008801571
#>  [6,] 0.148077480 0.133693757 0.148077480
#>  [7,] 0.036600725 0.036600725 0.036600725
#>  [8,] 0.099270395 0.285097107 0.099270395
#>  [9,] 0.013858041 0.013858041 0.013858041
#> [10,] 0.363562560 0.462260695 0.229926917
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.3204654 0.49172701 0.49172701
#>  [2,] 0.5728944 0.57289440 0.52094139
#>  [3,] 0.2177005 0.54393257 0.51007071
#>  [4,] 0.1709213 0.22013197 0.22013197
#>  [5,] 1.0797828 0.49171183 0.74452739
#>  [6,] 0.3366551 0.05528864 0.05528864
#>  [7,] 0.1295949 0.39984308 0.37872755
#>  [8,] 0.2004793 0.21506451 0.32008111
#>  [9,] 0.4865929 0.32182524 0.94259379
#> [10,] 0.5680805 0.16818576 0.23594006

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.077166020 0.07716602 0.07716602
#>  [2,] 0.290195372 0.39748889 0.15188302
#>  [3,] 0.098689994 0.06318855 0.48843305
#>  [4,] 0.205359641 0.08697949 0.33764527
#>  [5,] 0.310214509 0.31021451 0.31021451
#>  [6,] 0.194384502 0.19438450 0.19438450
#>  [7,] 0.002655225 0.55268784 0.24845983
#>  [8,] 0.057928993 0.07935155 0.35677746
#>  [9,] 0.587618892 0.58503139 0.94812641
#> [10,] 0.099377847 0.09937785 0.09937785
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.22194156 0.02176390 0.02176390
#>  [2,] 0.24813235 0.24813235 0.05572173
#>  [3,] 0.80385720 0.03690187 0.88179912
#>  [4,] 0.67542037 0.12407099 0.91429586
#>  [5,] 0.82028128 0.01834558 0.17136466
#>  [6,] 0.91912678 0.46269728 0.11815146
#>  [7,] 0.07413395 0.33086750 0.33086750
#>  [8,] 0.47588964 0.47588964 0.15694085
#>  [9,] 0.64148149 0.01258213 0.90726476
#> [10,] 0.75442729 0.80530579 0.52026835
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]      [,2]         [,3]
#>  [1,] 0.21836006 0.1736955 0.0003210508
#>  [2,] 0.73084219 0.7308422 0.3095542806
#>  [3,] 0.14821609 0.1482161 0.1482160939
#>  [4,] 0.10642777 0.1064278 0.7640100959
#>  [5,] 1.70604539 0.1834141 0.8516246777
#>  [6,] 0.83314856 0.9581327 0.1907602363
#>  [7,] 0.88584172 0.4982552 0.4982552099
#>  [8,] 0.08422768 0.2571781 0.0842276805
#>  [9,] 0.41589287 0.4158929 0.2322727559
#> [10,] 0.68446293 1.0906156 1.0036290766
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>               [,1]         [,2]         [,3]
#>  [1,] 6.101971e-02 7.722571e-01 6.590064e-01
#>  [2,] 1.146082e-02 2.914122e-01 1.146082e-02
#>  [3,] 2.801664e-02 2.801664e-02 3.692780e-02
#>  [4,] 4.616662e-01 7.101941e-01 4.938569e-01
#>  [5,] 9.612188e-05 9.612188e-05 9.612188e-05
#>  [6,] 1.248552e-01 1.248552e-01 1.069597e-02
#>  [7,] 9.899922e-02 9.039248e-02 9.899922e-02
#>  [8,] 7.819926e-01 7.819926e-01 1.538162e-01
#>  [9,] 9.603032e-02 9.603032e-02 9.603032e-02
#> [10,] 1.548518e-01 1.548518e-01 2.756923e-01

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27877215 0.27877215 0.27877215
#>  [2,] 0.02713961 0.30743099 0.32532796
#>  [3,] 1.16897350 0.40246803 2.13565647
#>  [4,] 0.44701206 0.73970382 0.18615530
#>  [5,] 0.53761670 1.68259994 0.49629911
#>  [6,] 0.01040477 0.01040477 0.01040477
#>  [7,] 0.25204344 0.25204344 0.31613573
#>  [8,] 0.37426065 0.67739677 0.26760110
#>  [9,] 0.18067095 0.18067095 0.18067095
#> [10,] 0.09479041 0.75849367 0.75849367
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.149474495 0.042865777 0.003982665
#>  [2,] 1.076083530 0.113451385 0.370615127
#>  [3,] 0.800930498 0.800930498 0.800930498
#>  [4,] 0.241509740 0.241509740 0.241509740
#>  [5,] 0.142315399 0.149693008 0.149693008
#>  [6,] 0.101138221 0.183540860 0.101138221
#>  [7,] 0.025687369 0.344964403 0.435475577
#>  [8,] 0.003066255 0.126458815 0.253988924
#>  [9,] 0.045074046 0.284127352 0.094188711
#> [10,] 0.002679729 0.002679729 0.002679729
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.07761162 0.26302956 0.04768799
#>  [2,] 0.07176797 0.07176797 0.17280733
#>  [3,] 0.62566816 0.39681283 0.14212395
#>  [4,] 0.39384877 0.66107622 0.39384877
#>  [5,] 0.28052138 0.28052138 0.26858016
#>  [6,] 1.00855626 0.29150281 0.36299734
#>  [7,] 0.07631658 0.03559139 0.20404511
#>  [8,] 0.02278319 0.67641287 0.33532672
#>  [9,] 0.33829285 0.33829285 0.08043957
#> [10,] 0.09493692 0.31046606 0.31046606
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.58090482 0.52825198 0.58090482
#>  [2,] 0.21174830 0.21174830 0.21174830
#>  [3,] 0.28832117 0.03106826 0.03106826
#>  [4,] 0.29115653 0.10804501 0.80407706
#>  [5,] 0.70451203 0.49083986 0.83064693
#>  [6,] 0.36539593 0.36539593 0.35134508
#>  [7,] 0.52837699 0.52182088 0.52837699
#>  [8,] 0.24113626 0.27335167 0.24395179
#>  [9,] 1.64920680 0.46746061 0.25033441
#> [10,] 0.02483839 0.02483839 0.02483839