Simulate from parametrized families of extendible MO distributions

Draws n iid d-variate samples from a parametrized family of extendible MO distributions.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

An integer for the number of samples.

d

An integer for the dimension.

a

A non-negative double representing the killing-rate \(a\) of the Bernstein function.

b

A non-negative double representing the drift \(b\) of the Bernstein function.

gamma

a positive double representing the scaling factor of the integral part of the Bernstein function.

eta

A numeric vector representing the distribution family's parameters, see the Details section.

family

A string representing the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family, see the Details section.

method

A string representing which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy–frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up to dimension \(30\).

Value

rpextmo returns a numeric matrix of size n x d. Each row corresponds to an independently and identically (iid) distributed sample from a d-variate parametrized extendible Marshall–Olkin distribution with the specified parameters.

Details

A parametrized ext. MO distribution is a family of ext. MO distributions, see rextmo(), corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, b \geq 0\). The \(\nu\) and \(\sigma\) represent the Lévy measure and Stieltjes measure, respectively. At least one of the following conditions must hold: \(a > 0\), \(b > 0\), or \(\nu \not\equiv 0\) (resp. \(\sigma \not\equiv 0\)).

Families

All implemented families are listed in the following; some re-combinations are possible, see ScaledBernsteinFunction, SumOfBernsteinFunctions, and CompositeScaledBernsteinFunction.

Armageddon family: We have for \(\nu = \sigma \equiv 0\) the Bernstein function \(\psi\): $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family: We have for \(\eta > 0\) the Bernstein function \(\psi\):: $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ with (discrete) Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family: We have \(\eta \in \mathbb{R}^2\) with \(\eta_1 \in {(0, 1)}, \eta_2 > 0\) a Bernstein function with Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family: We have for \(\eta > 0\) the Bernstein function \(\psi\): $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ with Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and with (discrete) Stieltjes measure \(\sigma\): $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family: We have for \(\eta \in {(0, 1)}\) the Bernstein function \(\psi\): $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ with Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and with Stieljtes measure \(\sigma\): $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family: We have for \(\eta > 0\) the Bernstein function \(\psi\): $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ with Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and with Stieltjes measure \(\sigma\): $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}} {u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family: We have for \(\eta > 0\) the Bernstein function \(\psi\): $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ with Lévy measure \(\nu\): $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and with Stieljtes measure \(\sigma\): $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used. For this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock-times.

The Lévy-frailty model (LFM) simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, n is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.89621129 2.0842032 3.5381627
#>  [2,] 1.03257499 2.9340503 2.0705893
#>  [3,] 0.36916084 3.9920367 1.4529960
#>  [4,] 3.78523642 3.1333137 0.6449898
#>  [5,] 0.16369341 4.4059409 0.5564593
#>  [6,] 0.46630046 0.1476089 0.4663005
#>  [7,] 0.04938859 0.8269391 0.8269391
#>  [8,] 0.36539716 0.3653972 0.3653972
#>  [9,] 0.73228974 1.9320894 2.9286017
#> [10,] 1.89723765 0.7176901 1.5282237
## comonotone
rpextmo(10, 3, a = 1)
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6810140 1.6810140 1.6810140
#>  [2,] 1.8425211 1.8425211 1.8425211
#>  [3,] 0.1213155 0.1213155 0.1213155
#>  [4,] 0.2720825 0.2720825 0.2720825
#>  [5,] 1.8494939 1.8494939 1.8494939
#>  [6,] 2.5877426 2.5877426 2.5877426
#>  [7,] 0.6417312 0.6417312 0.6417312
#>  [8,] 1.0822153 1.0822153 1.0822153
#>  [9,] 0.6904121 0.6904121 0.6904121
#> [10,] 0.6157218 0.6157218 0.6157218
## independence
rpextmo(10, 3, b = 1)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.40709804 1.48784761 1.6796328
#>  [2,] 0.79247883 0.20959998 2.4228128
#>  [3,] 1.28355319 3.34977485 1.1788727
#>  [4,] 0.84327665 0.05853796 4.2588416
#>  [5,] 2.43531709 0.33539233 0.3266379
#>  [6,] 0.51265654 0.15976861 0.1106287
#>  [7,] 3.20715497 1.37439915 0.4793136
#>  [8,] 0.02704653 1.66477329 0.2841166
#>  [9,] 0.66486077 0.09924463 3.3360172
#> [10,] 0.83610770 0.31648462 0.2765661

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.44279183 0.4427918 0.4427918
#>  [2,] 1.92128880 0.5874270 0.5820199
#>  [3,] 0.19383557 1.4004942 0.5825797
#>  [4,] 0.50712328 0.1380126 2.6428741
#>  [5,] 1.32982008 0.5798475 1.3298201
#>  [6,] 0.03437650 0.0343765 0.0343765
#>  [7,] 1.71119306 1.7111931 1.6354009
#>  [8,] 0.06703615 0.7992335 1.1912454
#>  [9,] 0.77550368 0.9367431 0.5257273
#> [10,] 1.75218760 1.2486361 0.8867321
## comonotone
rpextmo(10, 3, a = 1, method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06609500 0.06609500 0.06609500
#>  [2,] 0.24854872 0.24854872 0.24854872
#>  [3,] 0.56347715 0.56347715 0.56347715
#>  [4,] 0.56796047 0.56796047 0.56796047
#>  [5,] 1.23500528 1.23500528 1.23500528
#>  [6,] 1.62235398 1.62235398 1.62235398
#>  [7,] 0.04455849 0.04455849 0.04455849
#>  [8,] 0.36619181 0.36619181 0.36619181
#>  [9,] 0.06420805 0.06420805 0.06420805
#> [10,] 0.01388780 0.01388780 0.01388780
## independence
rpextmo(10, 3, b = 1, method = "ESM")
#>              [,1]       [,2]      [,3]
#>  [1,] 0.410020125 0.04331954 0.3709553
#>  [2,] 0.205320206 0.06678018 0.8461123
#>  [3,] 0.682522581 2.76714979 1.1642519
#>  [4,] 0.430645244 0.72283038 0.3090019
#>  [5,] 1.797598664 0.63018201 0.5841694
#>  [6,] 0.004310967 0.40139910 1.9675728
#>  [7,] 0.688441569 0.44523196 0.3219563
#>  [8,] 1.562508828 0.41686504 7.0614505
#>  [9,] 0.872413370 1.82506307 0.8467622
#> [10,] 0.077189736 0.78141338 2.1288718

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.72836470 1.3655610 0.5295375
#>  [2,] 2.12854818 1.1311362 0.6668655
#>  [3,] 0.54464555 1.2432002 2.4012770
#>  [4,] 2.48488732 0.3806837 4.1463513
#>  [5,] 2.03852740 0.4074661 1.0501524
#>  [6,] 0.19850696 2.4604301 2.8931309
#>  [7,] 4.27897311 1.6004939 2.4918548
#>  [8,] 0.06838816 0.0812121 0.0812121
#>  [9,] 5.16862432 0.8884238 0.4066343
#> [10,] 2.67035647 0.9486983 0.6333626
## comonotone
rpextmo(10, 3, a = 1, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.85785574 2.85785574 2.85785574
#>  [2,] 0.14855876 0.14855876 0.14855876
#>  [3,] 0.67868470 0.67868470 0.67868470
#>  [4,] 0.21024986 0.21024986 0.21024986
#>  [5,] 0.39111814 0.39111814 0.39111814
#>  [6,] 0.41757003 0.41757003 0.41757003
#>  [7,] 0.07291429 0.07291429 0.07291429
#>  [8,] 1.26816524 1.26816524 1.26816524
#>  [9,] 0.27908496 0.27908496 0.27908496
#> [10,] 0.80037174 0.80037174 0.80037174
## independence
rpextmo(10, 3, b = 1, method = "LFM")
#>              [,1]       [,2]      [,3]
#>  [1,] 0.140048902 0.93066662 0.3364388
#>  [2,] 1.877347645 3.23646556 0.6203077
#>  [3,] 0.261899964 0.27426001 0.8709385
#>  [4,] 0.039781756 0.69249110 0.5967041
#>  [5,] 1.647165636 1.24945564 0.8416705
#>  [6,] 0.001319481 0.07552047 0.8585607
#>  [7,] 0.782727998 0.77378015 0.4466014
#>  [8,] 0.859113725 0.04774431 2.8735176
#>  [9,] 0.676989209 0.03677058 1.4296343
#> [10,] 2.768762937 0.18183889 0.1060367

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.87835914 1.4939835 1.93689103
#>  [2,] 0.53299105 0.5329910 0.53299105
#>  [3,] 0.26964452 0.2696445 0.26964452
#>  [4,] 0.37239988 0.2239705 0.56762857
#>  [5,] 1.48524870 0.6794219 3.19817511
#>  [6,] 0.01622269 3.0880936 1.80446981
#>  [7,] 1.69519184 0.2585454 1.73541402
#>  [8,] 2.74275290 0.4062394 0.04972496
#>  [9,] 0.44622098 0.1905913 2.29717732
#> [10,] 0.08117781 0.0669228 0.08117781
## comonotone
rpextmo(10, 3, a = 1, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05128906 0.05128906 0.05128906
#>  [2,] 0.42494977 0.42494977 0.42494977
#>  [3,] 1.33246982 1.33246982 1.33246982
#>  [4,] 0.35864001 0.35864001 0.35864001
#>  [5,] 0.39001179 0.39001179 0.39001179
#>  [6,] 0.09554273 0.09554273 0.09554273
#>  [7,] 2.92553794 2.92553794 2.92553794
#>  [8,] 2.77771333 2.77771333 2.77771333
#>  [9,] 0.18316129 0.18316129 0.18316129
#> [10,] 0.26645069 0.26645069 0.26645069
## independence
rpextmo(10, 3, b = 1, method = "MDCM")
#>             [,1]       [,2]        [,3]
#>  [1,] 1.47709330 0.06463654 2.431923410
#>  [2,] 0.48466554 0.96624072 0.358047509
#>  [3,] 0.61100795 0.15042516 0.872952231
#>  [4,] 0.07760456 0.64669369 0.651835018
#>  [5,] 1.65207907 0.86603251 0.797382917
#>  [6,] 1.05694590 0.41452114 0.651099972
#>  [7,] 0.34495765 1.33725389 1.252801074
#>  [8,] 0.25202976 1.92940270 0.003086432
#>  [9,] 1.38411975 2.56640373 3.573963291
#> [10,] 0.80124656 0.70706268 1.812912002

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.99731826 0.48430071 0.51211736
#>  [2,] 0.33018979 0.34744121 2.32650335
#>  [3,] 1.73017814 1.00249434 1.03765527
#>  [4,] 0.61919409 0.61919409 0.05294610
#>  [5,] 0.02486854 0.02486854 0.02486854
#>  [6,] 0.42842286 1.79151058 2.02024563
#>  [7,] 1.72769900 0.82994022 0.94476401
#>  [8,] 2.00154015 0.65501744 2.54994633
#>  [9,] 1.61521986 0.08758411 0.29520690
#> [10,] 0.70152047 2.40248107 0.46042432
## comonotone
rpextmo(10, 3, a = 1, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.42325263 0.42325263 0.42325263
#>  [2,] 1.25672429 1.25672429 1.25672429
#>  [3,] 0.05332070 0.05332070 0.05332070
#>  [4,] 3.56592615 3.56592615 3.56592615
#>  [5,] 0.03112095 0.03112095 0.03112095
#>  [6,] 0.04975225 0.04975225 0.04975225
#>  [7,] 0.83653986 0.83653986 0.83653986
#>  [8,] 0.31289721 0.31289721 0.31289721
#>  [9,] 3.59696330 3.59696330 3.59696330
#> [10,] 0.27430244 0.27430244 0.27430244
## independence
rpextmo(10, 3, b = 1, method = "AM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.8360553 1.0123259 0.4097994
#>  [2,] 0.3994600 0.7153279 0.2680559
#>  [3,] 0.7935011 2.2642746 0.2462195
#>  [4,] 0.5856387 1.8867295 0.4583639
#>  [5,] 1.4802686 0.7811555 0.3359149
#>  [6,] 0.8572979 3.4626641 0.4291347
#>  [7,] 0.8481199 0.7522090 0.5962648
#>  [8,] 0.6882408 0.5434623 0.1053545
#>  [9,] 0.9922098 1.7011508 0.4205175
#> [10,] 0.4732692 0.1323255 0.9706242

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>              [,1]      [,2]      [,3]
#>  [1,] 1.938126465 1.9243528 0.9147238
#>  [2,] 1.121244932 1.6721203 1.8819600
#>  [3,] 1.575278642 0.5305009 1.3326253
#>  [4,] 1.305489952 3.4483681 5.6888674
#>  [5,] 1.443910940 0.2280847 1.4752116
#>  [6,] 2.892496765 5.9725644 0.3676216
#>  [7,] 0.004127329 0.9228158 0.9228158
#>  [8,] 3.551666681 0.6498255 1.0761295
#>  [9,] 1.070196558 2.1955393 1.7742697
#> [10,] 1.527591435 1.7552035 1.1302744
## comonotone
rpextmo(10, 3, a = 1, family = "Armageddon")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.38369868 0.38369868 0.38369868
#>  [2,] 0.07666186 0.07666186 0.07666186
#>  [3,] 0.02913211 0.02913211 0.02913211
#>  [4,] 3.24173555 3.24173555 3.24173555
#>  [5,] 2.10357454 2.10357454 2.10357454
#>  [6,] 0.03046840 0.03046840 0.03046840
#>  [7,] 1.00971579 1.00971579 1.00971579
#>  [8,] 0.18552788 0.18552788 0.18552788
#>  [9,] 2.99151403 2.99151403 2.99151403
#> [10,] 0.69290116 0.69290116 0.69290116
## independence
rpextmo(10, 3, b = 1, family = "Armageddon")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.41007998 0.3331954 0.4400782
#>  [2,] 0.79727230 0.4089764 0.9641383
#>  [3,] 0.22479230 1.1814492 2.1860416
#>  [4,] 2.19326550 0.1048520 1.6636980
#>  [5,] 0.06268378 0.1703145 1.8700177
#>  [6,] 0.03039613 0.8809782 1.0271963
#>  [7,] 0.44068440 0.8526026 2.7325575
#>  [8,] 3.17224009 0.6282484 0.8513373
#>  [9,] 1.86477836 0.5644167 0.4208404
#> [10,] 0.32612638 2.2733621 0.2577092

rpextmo(
  10, 3,
  a = 0.2, b = 0.5,
  family = "Armageddon",
  method = "ESM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 2.4584340 4.8565815 0.9577340
#>  [2,] 1.2918123 1.7651598 0.3664662
#>  [3,] 0.1232499 5.7654812 0.3070638
#>  [4,] 1.8542195 0.8257255 5.7491443
#>  [5,] 0.3874206 0.2830174 1.7798608
#>  [6,] 1.7728313 0.8521483 2.6032440
#>  [7,] 0.2773409 1.3151333 1.3151333
#>  [8,] 1.2231892 3.0099358 1.2189191
#>  [9,] 4.4190728 4.1856947 0.2694015
#> [10,] 0.6541937 1.6159650 1.4153201
## comonotone
rpextmo(
  10, 3,
  a = 1,
  family = "Armageddon",
  method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.96968439 1.96968439 1.96968439
#>  [2,] 0.70810874 0.70810874 0.70810874
#>  [3,] 0.44460351 0.44460351 0.44460351
#>  [4,] 0.09020110 0.09020110 0.09020110
#>  [5,] 0.49538333 0.49538333 0.49538333
#>  [6,] 0.65429407 0.65429407 0.65429407
#>  [7,] 0.07764398 0.07764398 0.07764398
#>  [8,] 0.91860524 0.91860524 0.91860524
#>  [9,] 0.26364057 0.26364057 0.26364057
#> [10,] 1.19124341 1.19124341 1.19124341
## independence
rpextmo(
  10, 3,
  b = 1,
  family = "Armageddon",
  method = "ESM"
)
#>             [,1]        [,2]       [,3]
#>  [1,] 0.01212768 0.313966308 0.62584544
#>  [2,] 0.23253937 0.008245956 0.10120260
#>  [3,] 0.26985475 0.127790034 1.83088918
#>  [4,] 0.42559192 1.807410676 0.06040545
#>  [5,] 1.11281792 0.265425964 0.01650703
#>  [6,] 1.66040691 0.095689961 0.52367275
#>  [7,] 0.70943021 0.078578082 0.88450787
#>  [8,] 0.12280007 1.219337656 1.94511170
#>  [9,] 0.78709395 0.260043557 0.24206683
#> [10,] 0.99797200 0.419092362 0.24887374

rpextmo(
  10, 3,
  a = 0.2, b = 0.5,
  family = "Armageddon",
  method = "LFM"
)
#>             [,1]      [,2]      [,3]
#>  [1,]  0.2623618 1.4259200 5.9235331
#>  [2,] 10.0044817 1.9863105 2.2330708
#>  [3,]  0.3750107 0.4012815 1.5768078
#>  [4,]  2.0585510 0.4769063 2.2237384
#>  [5,]  1.8239505 0.1705908 1.8239505
#>  [6,]  0.3031992 1.1851955 0.3906500
#>  [7,]  3.4403534 1.5937795 3.1091176
#>  [8,]  2.7349971 1.8138409 0.7851071
#>  [9,]  4.8139849 1.3360238 3.3062797
#> [10,]  0.4736677 0.6177787 1.4483524
## comonotone
rpextmo(
  10, 3,
  a = 1,
  family = "Armageddon",
  method = "LFM"
)
#>               [,1]         [,2]         [,3]
#>  [1,] 0.3261063732 0.3261063732 0.3261063732
#>  [2,] 0.5885375543 0.5885375543 0.5885375543
#>  [3,] 0.4048283181 0.4048283181 0.4048283181
#>  [4,] 0.7828417932 0.7828417932 0.7828417932
#>  [5,] 0.5653134762 0.5653134762 0.5653134762
#>  [6,] 0.0003980658 0.0003980658 0.0003980658
#>  [7,] 0.6126912632 0.6126912632 0.6126912632
#>  [8,] 1.8013075493 1.8013075493 1.8013075493
#>  [9,] 0.3451306480 0.3451306480 0.3451306480
#> [10,] 0.6913057356 0.6913057356 0.6913057356
## independence
rpextmo(
  10, 3,
  b = 1,
  family = "Armageddon",
  method = "LFM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.5064879 0.46096396 0.7761338
#>  [2,] 1.3147256 1.14686513 2.9144251
#>  [3,] 0.3502860 1.29326120 0.2692002
#>  [4,] 1.2960685 3.79347970 1.9559479
#>  [5,] 0.9365257 0.02338926 0.9443337
#>  [6,] 2.4498252 2.55495762 1.4829420
#>  [7,] 0.3530585 1.34039142 0.2882422
#>  [8,] 0.9268401 0.58472904 1.4550046
#>  [9,] 1.0242157 0.63676018 0.2824616
#> [10,] 0.6004604 0.17248792 0.1654850

rpextmo(
  10, 3,
  a = 0.2, b = 0.5,
  family = "Armageddon",
  method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.33633072 0.9515362 0.1474908
#>  [2,] 0.09307647 2.2159436 2.4065549
#>  [3,] 1.03599917 2.4532267 3.5005488
#>  [4,] 1.98080630 0.8288638 2.6458166
#>  [5,] 0.87010114 0.9907857 2.8378797
#>  [6,] 1.13053095 3.1322106 0.0400049
#>  [7,] 1.28504169 0.7292452 0.5127503
#>  [8,] 2.14758773 1.1431983 0.0473579
#>  [9,] 0.34804635 1.0337113 0.3466671
#> [10,] 0.85696867 1.1195955 1.1195955
## comonotone
rpextmo(
  10, 3,
  a = 1,
  family = "Armageddon",
  method = "MDCM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.737199530 1.737199530 1.737199530
#>  [2,] 2.606278631 2.606278631 2.606278631
#>  [3,] 0.467994520 0.467994520 0.467994520
#>  [4,] 0.737812204 0.737812204 0.737812204
#>  [5,] 0.719131352 0.719131352 0.719131352
#>  [6,] 0.125687358 0.125687358 0.125687358
#>  [7,] 2.747122092 2.747122092 2.747122092
#>  [8,] 1.190380410 1.190380410 1.190380410
#>  [9,] 0.244921049 0.244921049 0.244921049
#> [10,] 0.008889997 0.008889997 0.008889997
## independence
rpextmo(
  10, 3,
  b = 1,
  family = "Armageddon",
  method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.35015922 1.3793892 0.2135113
#>  [2,] 1.16954278 0.4925328 0.5895984
#>  [3,] 0.26141054 1.9473061 0.7216814
#>  [4,] 0.80181342 0.2366670 0.1021136
#>  [5,] 0.02805716 1.0543020 0.4091314
#>  [6,] 0.64303190 0.9055483 1.6370541
#>  [7,] 1.12274308 0.7486118 0.0706893
#>  [8,] 0.89268619 3.4277502 2.1333087
#>  [9,] 0.74863393 0.1834828 1.2208946
#> [10,] 0.31474647 1.9647424 0.6395613

rpextmo(
  10, 3,
  a = 0.2, b = 0.5,
  family = "Armageddon",
  method = "AM"
)
#>              [,1]      [,2]      [,3]
#>  [1,] 1.916990631 1.9775960 2.2351090
#>  [2,] 2.578441508 1.9084050 2.6844572
#>  [3,] 1.183363664 1.1833637 1.1833637
#>  [4,] 0.369018350 1.2415112 1.9975318
#>  [5,] 0.566798524 0.5667985 0.5667985
#>  [6,] 0.005228992 1.1371434 1.0338168
#>  [7,] 1.283549956 1.1347429 1.2835500
#>  [8,] 0.675511026 0.1868192 1.0689827
#>  [9,] 3.733589028 4.3659582 0.9979497
#> [10,] 0.042802219 2.7256658 1.1556354
## comonotone
rpextmo(
  10, 3,
  a = 1,
  family = "Armageddon",
  method = "AM"
)
#>              [,1]        [,2]        [,3]
#>  [1,] 1.321498486 1.321498486 1.321498486
#>  [2,] 0.594876691 0.594876691 0.594876691
#>  [3,] 2.464490784 2.464490784 2.464490784
#>  [4,] 0.547399436 0.547399436 0.547399436
#>  [5,] 0.165612901 0.165612901 0.165612901
#>  [6,] 0.050233243 0.050233243 0.050233243
#>  [7,] 1.812641806 1.812641806 1.812641806
#>  [8,] 0.009362767 0.009362767 0.009362767
#>  [9,] 2.044736134 2.044736134 2.044736134
#> [10,] 0.215938923 0.215938923 0.215938923
## independence
rpextmo(
  10, 3,
  b = 1,
  family = "Armageddon",
  method = "AM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.99587371 0.03714099 1.34252627
#>  [2,] 0.16174384 2.53313655 0.55171533
#>  [3,] 0.53650718 0.03638600 0.52934046
#>  [4,] 0.32005828 0.58427340 0.37710530
#>  [5,] 0.39184240 2.90868886 0.01428954
#>  [6,] 1.18160221 1.13506756 1.03067598
#>  [7,] 0.12636168 1.39877799 1.31931319
#>  [8,] 0.02372679 0.57788295 2.30416957
#>  [9,] 1.41718399 0.31674070 1.12204694
#> [10,] 0.98192886 1.32971664 0.45860550

## Poisson
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Poisson"
)
#>            [,1]       [,2]       [,3]
#>  [1,] 0.5384901 2.94813208 0.53849015
#>  [2,] 0.1017191 0.20041889 0.20041889
#>  [3,] 0.5203302 1.68125615 1.68125615
#>  [4,] 0.6172509 0.07564676 0.49279121
#>  [5,] 0.4134294 0.41342935 0.17597587
#>  [6,] 1.1171599 0.52632538 1.66544876
#>  [7,] 0.1135298 0.11352976 0.11352976
#>  [8,] 1.3522320 0.17712849 0.17712849
#>  [9,] 0.6229397 0.20377457 0.20377457
#> [10,] 0.4708807 0.47088074 0.05042807
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Poisson",
  method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 2.43079020 1.17028967 1.32820619
#>  [2,] 0.12397769 0.52013540 0.52013540
#>  [3,] 0.09643320 0.36050359 0.36050359
#>  [4,] 0.11645007 1.36989266 0.14540044
#>  [5,] 0.52898318 0.48308691 0.29063290
#>  [6,] 0.82812959 0.82812959 0.82812959
#>  [7,] 0.74103510 0.02667862 1.12138619
#>  [8,] 0.95942351 0.50026351 0.44200090
#>  [9,] 0.09438203 0.26381668 0.08334398
#> [10,] 0.31933792 0.49854279 0.78398464
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Poisson",
  method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.28442456 1.61863272 0.86589514
#>  [2,] 0.09925539 0.09925539 0.09925539
#>  [3,] 0.47298241 1.91588460 0.18299160
#>  [4,] 0.23050878 0.19900228 1.22240209
#>  [5,] 0.03472681 0.23119265 0.78710040
#>  [6,] 0.27645898 0.38021820 0.27645898
#>  [7,] 0.11729923 0.11729923 0.11729923
#>  [8,] 0.11653738 0.09861330 1.48931647
#>  [9,] 1.64231168 0.81155273 0.95508818
#> [10,] 0.43238588 1.09517192 0.30073282
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Poisson",
  method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.83163425 0.39332444 1.65798683
#>  [2,] 0.08378116 0.05957384 0.05957384
#>  [3,] 0.65562064 0.28198679 0.23941361
#>  [4,] 0.64311122 0.05109943 0.64311122
#>  [5,] 0.21249733 0.11958074 0.11958074
#>  [6,] 1.35980242 0.01116162 0.01116162
#>  [7,] 0.20884014 0.20884014 0.20884014
#>  [8,] 0.10716782 0.78009466 0.45235794
#>  [9,] 0.33656941 0.40392144 0.28408869
#> [10,] 2.24581197 0.27656755 0.27656755
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Poisson",
  method = "AM"
)
#>              [,1]       [,2]      [,3]
#>  [1,] 0.001350696 0.60377575 0.6037758
#>  [2,] 1.195928400 0.39072463 1.4858236
#>  [3,] 1.169574206 0.82208132 0.4785511
#>  [4,] 0.879878990 0.36450902 0.8798790
#>  [5,] 0.087564714 0.06219288 0.7890873
#>  [6,] 0.504653993 0.19882999 0.1880001
#>  [7,] 4.171635567 0.34309057 0.3430906
#>  [8,] 0.600307222 0.46027151 0.8851565
#>  [9,] 1.511836454 0.21182177 0.6172429
#> [10,] 0.829267113 0.26551310 0.4296528

## Pareto
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = c(0.5, 1e-4), family = "Pareto"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.03801943 0.65604880 0.6560488
#>  [2,] 3.60108700 0.74417377 2.8460586
#>  [3,] 0.31922982 4.52410839 0.2823101
#>  [4,] 1.27952712 1.06636982 0.4914272
#>  [5,] 2.42398714 3.46157273 0.6318315
#>  [6,] 0.17385777 0.89405255 0.8940526
#>  [7,] 2.68123944 8.50681993 2.0407879
#>  [8,] 0.72334847 0.03523473 0.8039953
#>  [9,] 0.01001097 0.34361360 3.7353608
#> [10,] 1.89887487 1.08444176 0.6752454
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = c(0.5, 1e-4), family = "Pareto",
  method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.53043624 0.08917464 2.10482031
#>  [2,] 3.96575635 0.09047190 1.75178957
#>  [3,] 0.01329087 0.01329087 0.01329087
#>  [4,] 0.13509027 0.55130302 0.25220644
#>  [5,] 0.80385731 1.61296322 0.36377964
#>  [6,] 0.18037861 3.35820185 0.11371902
#>  [7,] 2.63125259 2.32816778 0.51317384
#>  [8,] 1.27527803 4.81382291 0.20352545
#>  [9,] 1.70279273 5.93255358 1.65909954
#> [10,] 3.03615947 0.48737702 5.12532961
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = c(0.5, 1e-4), family = "Pareto",
  method = "LFM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.3802050 0.2712353 0.9743875
#>  [2,] 1.7928790 1.9662166 4.0124992
#>  [3,] 1.2947774 0.6111844 1.2947774
#>  [4,] 0.3219877 0.4421034 3.5818796
#>  [5,] 0.2987168 0.2987168 0.2987168
#>  [6,] 3.5240904 0.1640650 0.1175563
#>  [7,] 1.1947760 2.0892588 1.0199826
#>  [8,] 0.6710051 0.1058966 0.9582827
#>  [9,] 0.2628965 1.2797715 2.1528078
#> [10,] 2.0511174 0.7190057 0.4311222
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = c(0.5, 1e-4), family = "Pareto",
  method = "MDCM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 1.78511648 1.7851165 1.7851165
#>  [2,] 0.45031283 0.8511236 0.4967267
#>  [3,] 2.25475203 2.3884482 2.8899155
#>  [4,] 0.06972239 4.0369886 0.1178982
#>  [5,] 1.35833505 3.8977334 0.4302498
#>  [6,] 1.29581633 2.5822523 5.4889835
#>  [7,] 0.75015457 1.2888028 1.2888028
#>  [8,] 0.78183056 0.9798557 0.6210200
#>  [9,] 3.49188211 0.5245863 1.3741688
#> [10,] 0.61538900 0.6153890 0.4316032
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = c(0.5, 1e-4), family = "Pareto",
  method = "AM"
)
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2046806 1.1587371 0.3899301
#>  [2,] 2.5762738 0.7269645 1.0138938
#>  [3,] 0.9149128 2.9388244 3.2952497
#>  [4,] 1.0490644 1.2558981 1.7355325
#>  [5,] 2.0638111 3.4190247 1.4019185
#>  [6,] 0.4646360 8.4839746 5.9659956
#>  [7,] 0.6213153 0.5319838 0.6213153
#>  [8,] 1.2663465 0.3985395 1.2663465
#>  [9,] 0.6243965 0.6243965 0.6243965
#> [10,] 0.2810049 0.2810049 0.2810049

## Exponential
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Exponential"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.76915167 0.16643010 0.76915167
#>  [2,] 0.72384121 0.72384121 0.72384121
#>  [3,] 0.27848983 0.82470260 0.63694234
#>  [4,] 0.03642472 0.03642472 0.52975088
#>  [5,] 0.48350055 0.48350055 0.21553313
#>  [6,] 0.11716964 0.44770357 0.26467111
#>  [7,] 0.76499150 0.76499150 0.14882730
#>  [8,] 0.25587822 0.25587822 0.33556853
#>  [9,] 0.33816101 0.33816101 0.33816101
#> [10,] 0.18633327 0.71142430 0.05604849
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Exponential",
  method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04045609 0.04045609 0.04045609
#>  [2,] 0.61948467 0.83121140 0.83121140
#>  [3,] 0.80668842 0.23160683 0.80668842
#>  [4,] 0.61069644 0.61069644 0.16759510
#>  [5,] 0.29933560 0.44578367 0.44578367
#>  [6,] 0.95513595 0.95513595 0.95513595
#>  [7,] 0.23405182 0.23405182 0.23405182
#>  [8,] 0.73397860 1.46113314 1.34970633
#>  [9,] 0.57296737 0.57296737 0.57296737
#> [10,] 0.39049138 0.12680916 0.39049138
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Exponential",
  method = "LFM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23483294 0.99321744 0.56400377
#>  [2,] 0.24431059 0.24431059 0.24431059
#>  [3,] 0.05830077 0.05830077 0.05830077
#>  [4,] 0.07863727 0.17004309 0.13542599
#>  [5,] 0.16102862 2.31560185 0.16102862
#>  [6,] 0.87523643 0.87523643 1.34311737
#>  [7,] 0.67926441 0.20844698 0.47263393
#>  [8,] 0.21356090 0.21356090 0.21356090
#>  [9,] 0.36785517 0.36785517 0.36785517
#> [10,] 0.31870396 0.39930528 0.31870396
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Exponential",
  method = "MDCM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.77928387 0.65551992 0.3285593
#>  [2,] 0.08902028 0.08477003 0.5310817
#>  [3,] 0.69517226 0.69517226 0.6705815
#>  [4,] 1.79767599 0.10844114 1.7001421
#>  [5,] 0.54178692 0.54178692 0.5417869
#>  [6,] 0.49309254 0.41700532 1.0259115
#>  [7,] 0.18092478 0.19794202 0.1979420
#>  [8,] 0.23914103 0.29003162 0.2900316
#>  [9,] 1.01615056 1.01615056 0.8888574
#> [10,] 0.06165271 1.39569173 0.6276128
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Exponential",
  method = "AM"
)
#>             [,1]       [,2]      [,3]
#>  [1,] 0.12815252 1.24027251 0.1281525
#>  [2,] 0.16768521 0.27263236 0.3655216
#>  [3,] 0.44073798 1.65687245 1.0118053
#>  [4,] 0.87220602 0.87220602 0.3562619
#>  [5,] 0.25566983 0.08195684 0.2556698
#>  [6,] 1.17582650 0.31296611 0.1271911
#>  [7,] 0.48832033 0.48832033 0.2283399
#>  [8,] 0.96462034 0.19951242 0.3631285
#>  [9,] 1.15618385 1.15618385 1.1561839
#> [10,] 0.03859964 0.03859964 0.0250658

## Alpha-Stable
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "AlphaStable"
)
#>            [,1]       [,2]        [,3]
#>  [1,] 0.1188554 0.05302391 0.412475456
#>  [2,] 0.7702091 0.77020913 0.770209134
#>  [3,] 0.0839063 0.81381170 0.813811703
#>  [4,] 0.7135108 0.25968423 0.462032316
#>  [5,] 0.5799051 0.36311545 0.002914865
#>  [6,] 0.3342043 0.13352202 0.133522024
#>  [7,] 0.2883869 0.28838686 0.282204630
#>  [8,] 0.3053661 0.23389301 0.107816288
#>  [9,] 0.3735397 0.19252756 0.373539692
#> [10,] 0.3457597 0.32886734 0.281045297
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "AlphaStable",
  method = "ESM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02329084 0.36554151 0.01330986
#>  [2,] 0.03838716 0.57700576 0.07444418
#>  [3,] 0.11456016 0.11456016 0.11456016
#>  [4,] 0.36022416 0.36022416 0.18677401
#>  [5,] 0.02891450 0.02891450 0.02891450
#>  [6,] 0.51510561 0.51510561 0.51510561
#>  [7,] 0.09445756 1.26698528 0.37814897
#>  [8,] 0.19040417 0.80059174 0.80059174
#>  [9,] 0.06970461 0.06970461 0.01265433
#> [10,] 0.13203102 0.13203102 0.13203102
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "AlphaStable",
  method = "MDCM"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.33351279 0.33351279 0.77372059
#>  [2,] 0.06575070 1.42587344 1.04525281
#>  [3,] 0.49704204 0.68955967 1.06825316
#>  [4,] 1.06634263 0.35679423 0.06227436
#>  [5,] 0.08653161 0.08653161 0.08653161
#>  [6,] 0.13629865 0.13629865 0.13629865
#>  [7,] 0.46694348 0.98643556 0.98643556
#>  [8,] 0.13447124 0.13447124 0.13447124
#>  [9,] 0.42445918 0.13062162 0.42445918
#> [10,] 0.87831931 0.43523953 0.19803887
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "AlphaStable",
  method = "AM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2465518 0.33863075 0.3386308
#>  [2,] 0.7129376 0.06938957 0.4317275
#>  [3,] 0.5436500 0.54364998 0.2107166
#>  [4,] 1.5905299 0.83358536 0.4957854
#>  [5,] 0.1238712 0.12387117 0.1238712
#>  [6,] 0.2917123 0.02932120 0.2917123
#>  [7,] 0.4784681 0.16577637 0.3562116
#>  [8,] 0.5136342 0.43021607 0.7039487
#>  [9,] 0.1376318 0.13561210 0.1376318
#> [10,] 0.6001202 0.60012019 0.6001202

## Inverse Gaussian
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "InverseGaussian"
)
#>            [,1]        [,2]       [,3]
#>  [1,] 0.1409705 0.211289726 0.03060706
#>  [2,] 0.1488620 0.036603626 0.14886203
#>  [3,] 0.5300980 0.078149027 0.07814903
#>  [4,] 0.1146304 0.138130417 0.21750741
#>  [5,] 1.5508637 0.203634722 0.13054224
#>  [6,] 0.7957410 0.081907590 0.64900377
#>  [7,] 0.0714812 0.071481203 0.07148120
#>  [8,] 0.2857447 0.285744735 0.23571288
#>  [9,] 0.4722825 0.221864859 0.57581621
#> [10,] 1.0141005 0.009328293 1.01410047
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "InverseGaussian",
  method = "ESM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.03947556 0.5076309 0.2523263
#>  [2,] 0.39769625 0.1831559 0.3976962
#>  [3,] 0.47185714 0.5146175 0.3523747
#>  [4,] 0.48027901 0.7211476 1.1257791
#>  [5,] 0.60891544 0.9399471 0.4173447
#>  [6,] 0.49484063 1.0237072 0.3588664
#>  [7,] 0.89644132 1.3189379 1.2627992
#>  [8,] 0.50962799 0.1060397 0.1060397
#>  [9,] 1.05725709 0.1130717 0.1130717
#> [10,] 0.80349980 0.1655493 0.1857617
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "InverseGaussian",
  method = "MDCM"
)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.0703101 0.07031010 0.0703101
#>  [2,] 0.2387464 0.20395290 0.2039529
#>  [3,] 0.5683093 0.56830934 0.2533325
#>  [4,] 0.1327893 0.13278930 0.1327893
#>  [5,] 0.1321343 0.13213433 0.1004981
#>  [6,] 0.3346876 0.33468763 0.3346876
#>  [7,] 0.2413177 0.10257467 0.2538688
#>  [8,] 0.2267799 0.11470839 0.2994309
#>  [9,] 0.3183641 0.31836409 0.6799437
#> [10,] 0.2623398 0.07352612 0.3944653
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "InverseGaussian",
  method = "AM"
)
#>             [,1]      [,2]      [,3]
#>  [1,] 0.13859723 0.2815087 0.2815087
#>  [2,] 0.10478917 0.2620443 0.2620443
#>  [3,] 0.35926681 0.1447893 0.3252922
#>  [4,] 0.54965322 0.5496532 0.5496532
#>  [5,] 1.39135469 0.4266410 0.2337817
#>  [6,] 0.09795349 1.5859128 0.7695346
#>  [7,] 0.18069316 0.6517452 0.6801263
#>  [8,] 0.50849916 0.0823517 0.0823517
#>  [9,] 0.01592301 0.4540579 0.2622479
#> [10,] 0.01442262 0.1188920 0.1188920

## Gamma
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Gamma"
)
#>             [,1]       [,2]       [,3]
#>  [1,] 0.87382362 0.45353225 0.11039472
#>  [2,] 0.52726451 0.49449616 0.49449616
#>  [3,] 0.24545930 0.61409310 0.21106819
#>  [4,] 0.46788755 0.31526734 1.07230665
#>  [5,] 0.01989555 0.01989555 0.01989555
#>  [6,] 0.14245913 0.14245913 0.14245913
#>  [7,] 0.44751464 0.93845651 0.96533447
#>  [8,] 1.00067653 0.56126711 0.02000618
#>  [9,] 0.31570383 0.28536390 0.31570383
#> [10,] 0.39225547 0.39225547 0.74795137
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Gamma",
  method = "ESM"
)
#>             [,1]       [,2]         [,3]
#>  [1,] 0.02195903 0.02195903 0.6202911504
#>  [2,] 0.14359179 0.02474621 0.6951709702
#>  [3,] 0.54628062 0.22019885 0.0699231768
#>  [4,] 1.56515099 0.88623286 0.5205399685
#>  [5,] 0.05512926 0.18137553 0.0551292628
#>  [6,] 0.46639727 0.15430361 0.4663972650
#>  [7,] 0.10585458 0.62975424 0.0002814472
#>  [8,] 0.51589526 0.35387612 0.6276545579
#>  [9,] 0.88477892 1.06255531 0.4094883338
#> [10,] 0.59449528 1.24403443 0.0276797126
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Gamma",
  method = "MDCM"
)
#>             [,1]       [,2]        [,3]
#>  [1,] 0.38593180 0.21614067 0.385931796
#>  [2,] 0.34382990 0.35204716 0.751211498
#>  [3,] 0.45927337 0.45927337 0.459273374
#>  [4,] 1.02377673 0.31909043 0.738474014
#>  [5,] 0.04818464 0.06456082 0.366821413
#>  [6,] 1.47420270 0.56093040 0.007862795
#>  [7,] 0.00722717 0.03645918 0.316703489
#>  [8,] 0.01644754 0.18701075 0.449882589
#>  [9,] 0.88866072 0.88866072 0.305909929
#> [10,] 0.57358661 0.70343452 0.703434517
rpextmo(
  10, 3,
  a = 0.2, b = 0.5, gamma = 2,
  eta = 0.5, family = "Gamma",
  method = "AM"
)
#>              [,1]       [,2]        [,3]
#>  [1,] 0.008780949 0.31487127 0.008780949
#>  [2,] 0.067630716 0.06763072 0.067630716
#>  [3,] 0.928175848 0.02174718 0.151552663
#>  [4,] 0.677361545 0.61376902 0.120702655
#>  [5,] 0.338373682 0.61301768 0.613017678
#>  [6,] 1.037612409 1.03761241 0.186373869
#>  [7,] 0.004625806 0.01531647 0.015316468
#>  [8,] 0.056476142 0.95261295 0.247298262
#>  [9,] 0.050948530 0.09176461 0.091764607
#> [10,] 0.241978999 0.17685520 0.241978999