Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.3989274 0.39892742 0.3989274
#>  [2,] 0.2267036 0.03532953 0.2267036
#>  [3,] 3.6117778 1.22199809 0.5002812
#>  [4,] 0.5647787 1.45407763 0.2906022
#>  [5,] 3.3863487 3.38634870 1.0424181
#>  [6,] 0.7761562 0.77615624 0.7761562
#>  [7,] 0.8940514 0.50486811 0.9144062
#>  [8,] 1.2524212 1.17915545 1.5052246
#>  [9,] 1.1732171 0.71315277 0.1410169
#> [10,] 0.9870023 1.29097491 2.6910488
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17369504 0.17369504 0.17369504
#>  [2,] 2.09294152 2.09294152 2.09294152
#>  [3,] 0.55839158 0.55839158 0.55839158
#>  [4,] 0.12862352 0.12862352 0.12862352
#>  [5,] 1.01411500 1.01411500 1.01411500
#>  [6,] 1.15457242 1.15457242 1.15457242
#>  [7,] 0.03529932 0.03529932 0.03529932
#>  [8,] 0.87742091 0.87742091 0.87742091
#>  [9,] 0.02274232 0.02274232 0.02274232
#> [10,] 0.19200900 0.19200900 0.19200900
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]       [,2]         [,3]
#>  [1,] 1.89135936 0.03215321 1.1388865718
#>  [2,] 0.06131707 0.34663385 0.0004410759
#>  [3,] 0.48277774 0.06463915 2.7763090505
#>  [4,] 0.15429343 2.39500483 1.2448800671
#>  [5,] 1.21862395 0.48566011 1.7594123605
#>  [6,] 1.87545300 3.19568040 1.1322888322
#>  [7,] 0.87361273 0.67134800 2.0376409589
#>  [8,] 0.80811449 1.28491868 0.2340890315
#>  [9,] 0.05015572 0.02929445 0.5514568617
#> [10,] 0.71268862 0.10826424 0.9372899951

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2480360 0.02346271 0.2480360
#>  [2,] 1.4018904 2.28990894 2.2899089
#>  [3,] 0.5348709 0.47553858 0.9888667
#>  [4,] 0.3056650 0.07709115 0.3056650
#>  [5,] 0.6186045 0.55758073 0.3333249
#>  [6,] 1.0332223 0.36539192 0.5994917
#>  [7,] 0.7881750 0.10641435 3.2703888
#>  [8,] 1.0086597 1.08404407 0.1656118
#>  [9,] 2.6344532 1.17142380 2.4381955
#> [10,] 0.9085697 1.90001725 1.9795452
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4996765 0.4996765 0.4996765
#>  [2,] 0.9865551 0.9865551 0.9865551
#>  [3,] 0.5793855 0.5793855 0.5793855
#>  [4,] 0.3496005 0.3496005 0.3496005
#>  [5,] 0.9785055 0.9785055 0.9785055
#>  [6,] 0.4995919 0.4995919 0.4995919
#>  [7,] 2.0248645 2.0248645 2.0248645
#>  [8,] 2.7755527 2.7755527 2.7755527
#>  [9,] 0.3115411 0.3115411 0.3115411
#> [10,] 0.2188506 0.2188506 0.2188506
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.85969566 0.3230405 5.68930756
#>  [2,] 0.92378581 1.1084272 0.09124955
#>  [3,] 0.11292246 0.6451023 0.88354010
#>  [4,] 6.25281141 0.3751165 2.90379344
#>  [5,] 0.43285960 1.1385463 2.04026046
#>  [6,] 0.10097283 1.3132022 1.22001220
#>  [7,] 0.53871367 2.7812513 0.54399461
#>  [8,] 0.05654117 2.6156425 1.63388908
#>  [9,] 1.09957283 1.9328172 1.35808589
#> [10,] 0.90209291 0.1207490 0.28479271

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.44389475 0.4438947 0.1311923
#>  [2,] 1.82348842 1.0535268 2.3770010
#>  [3,] 0.99501061 1.4327405 1.5863887
#>  [4,] 0.03833941 2.3508759 1.9646983
#>  [5,] 1.50249733 3.5306900 1.0878357
#>  [6,] 0.92466197 4.7214150 0.8879158
#>  [7,] 0.17014589 3.6119634 0.9475345
#>  [8,] 0.86654271 0.8099894 1.7162572
#>  [9,] 1.08746206 2.1421241 3.0035997
#> [10,] 0.59060658 1.3122472 2.1590146
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 2.153151234 2.153151234 2.153151234
#>  [2,] 0.288168004 0.288168004 0.288168004
#>  [3,] 1.115001806 1.115001806 1.115001806
#>  [4,] 0.002287961 0.002287961 0.002287961
#>  [5,] 0.183878495 0.183878495 0.183878495
#>  [6,] 0.077691326 0.077691326 0.077691326
#>  [7,] 2.368304701 2.368304701 2.368304701
#>  [8,] 0.819211355 0.819211355 0.819211355
#>  [9,] 0.215342564 0.215342564 0.215342564
#> [10,] 1.172727856 1.172727856 1.172727856
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1194191 0.61051526 0.9852439
#>  [2,] 4.0348499 0.42208322 3.8501501
#>  [3,] 0.9438850 0.61468105 1.6001631
#>  [4,] 0.3764923 0.56899409 0.4831234
#>  [5,] 0.7545569 0.26980307 0.3442345
#>  [6,] 0.1026198 0.14846798 5.2607586
#>  [7,] 0.2296799 0.02157811 1.1393150
#>  [8,] 0.4045672 1.11732622 0.9120696
#>  [9,] 0.7081232 2.63865115 0.4315466
#> [10,] 0.5423724 0.55993308 1.0641757

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7333224 0.88512030 0.9173051
#>  [2,] 1.0706360 2.07123443 3.1276421
#>  [3,] 0.8417366 0.18788237 0.8417366
#>  [4,] 1.4448217 0.69710341 2.8948543
#>  [5,] 1.3917089 2.44515899 0.8120568
#>  [6,] 0.2026034 0.20260342 0.2026034
#>  [7,] 1.5723551 1.57235511 1.5723551
#>  [8,] 0.3523100 0.06412867 0.3523100
#>  [9,] 0.2003533 1.00610868 0.4677182
#> [10,] 0.6530671 0.80732638 0.9661515
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.11546734 0.11546734 0.11546734
#>  [2,] 0.47335788 0.47335788 0.47335788
#>  [3,] 0.62799113 0.62799113 0.62799113
#>  [4,] 0.23424346 0.23424346 0.23424346
#>  [5,] 1.86820008 1.86820008 1.86820008
#>  [6,] 0.93679127 0.93679127 0.93679127
#>  [7,] 0.02616229 0.02616229 0.02616229
#>  [8,] 0.44605221 0.44605221 0.44605221
#>  [9,] 0.13253314 0.13253314 0.13253314
#> [10,] 0.50137222 0.50137222 0.50137222
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 0.25657796 0.6503190 0.3343308
#>  [2,] 0.69826839 0.3711049 1.8350194
#>  [3,] 1.07890181 0.9595054 1.2244452
#>  [4,] 0.53671685 0.6369403 2.0700067
#>  [5,] 0.27239513 0.6872711 1.1794906
#>  [6,] 0.57959636 1.0564164 1.2899436
#>  [7,] 0.03324996 1.1676864 0.1687433
#>  [8,] 0.77729287 0.5679252 1.5997936
#>  [9,] 0.99018717 0.5414021 0.6818241
#> [10,] 1.22087662 1.1709513 0.5603087

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>           [,1]      [,2]      [,3]
#>  [1,] 1.699553 1.2940704 1.2745207
#>  [2,] 2.154019 1.7259224 5.0718004
#>  [3,] 2.411248 0.3963711 2.6395423
#>  [4,] 1.187820 0.4850115 2.4941135
#>  [5,] 1.944088 0.3925745 0.2254777
#>  [6,] 2.520776 0.3970377 0.7837829
#>  [7,] 1.651041 0.5716164 6.2111359
#>  [8,] 0.294503 0.2945030 0.2945030
#>  [9,] 3.135461 1.2923494 2.0298148
#> [10,] 1.522286 1.5222856 1.5222856
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.7951054 1.7951054 1.7951054
#>  [2,] 0.2168951 0.2168951 0.2168951
#>  [3,] 0.1562837 0.1562837 0.1562837
#>  [4,] 1.2250498 1.2250498 1.2250498
#>  [5,] 1.4959965 1.4959965 1.4959965
#>  [6,] 5.9877194 5.9877194 5.9877194
#>  [7,] 1.3057608 1.3057608 1.3057608
#>  [8,] 0.3143301 0.3143301 0.3143301
#>  [9,] 0.3115415 0.3115415 0.3115415
#> [10,] 0.7013091 0.7013091 0.7013091
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>            [,1]       [,2]        [,3]
#>  [1,] 0.6970753 0.82003203 0.218462836
#>  [2,] 2.6466460 2.10849512 2.092007234
#>  [3,] 0.6195920 0.16118378 1.042034198
#>  [4,] 0.6243589 0.78173963 0.056977913
#>  [5,] 0.8384499 0.01967884 0.596320380
#>  [6,] 1.0665317 1.81927760 0.187117333
#>  [7,] 1.8267594 0.00865753 0.012143698
#>  [8,] 0.2484481 0.71212631 0.205804161
#>  [9,] 0.6402306 1.19018840 1.461531671
#> [10,] 0.1400665 2.15942193 0.008295272

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.883997489 2.3976371 2.4527161
#>  [2,] 0.172218854 0.6484462 0.6484462
#>  [3,] 0.218000377 3.7813032 0.5518667
#>  [4,] 0.436950775 0.4369508 0.4369508
#>  [5,] 2.237879480 0.5702489 1.7659275
#>  [6,] 0.009944931 0.1684309 0.1589054
#>  [7,] 0.200384170 4.7229389 2.9205471
#>  [8,] 0.177221563 2.2347812 1.9500609
#>  [9,] 2.968862993 2.0541546 1.4684117
#> [10,] 0.068876466 0.4275214 0.7662788
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4836397 0.4836397 0.4836397
#>  [2,] 0.1074682 0.1074682 0.1074682
#>  [3,] 2.0723806 2.0723806 2.0723806
#>  [4,] 0.3751351 0.3751351 0.3751351
#>  [5,] 0.7076808 0.7076808 0.7076808
#>  [6,] 0.3217921 0.3217921 0.3217921
#>  [7,] 0.4440674 0.4440674 0.4440674
#>  [8,] 0.6060086 0.6060086 0.6060086
#>  [9,] 1.1391718 1.1391718 1.1391718
#> [10,] 0.4301081 0.4301081 0.4301081
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]        [,2]        [,3]
#>  [1,] 0.55211259 0.306947087 0.867022137
#>  [2,] 0.53879301 1.286239509 1.045400385
#>  [3,] 0.76315762 0.001541482 0.071932403
#>  [4,] 0.52164683 1.044270635 0.797797514
#>  [5,] 0.37892853 0.533932632 3.386667780
#>  [6,] 0.02780811 1.733115949 0.010440995
#>  [7,] 3.62442190 0.174089203 0.006619458
#>  [8,] 3.91018050 0.695505875 0.161907847
#>  [9,] 0.22450587 0.338770007 2.736910542
#> [10,] 0.33552588 0.699358729 0.332060338

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.0683081 0.2492426 0.41692294
#>  [2,] 0.8777274 0.8777274 0.87772737
#>  [3,] 0.4161337 0.4161337 0.41421777
#>  [4,] 2.9065960 0.8516188 1.82327412
#>  [5,] 3.1587040 4.2202497 0.06201419
#>  [6,] 1.8272500 3.2463627 0.73063133
#>  [7,] 1.3527236 1.3527236 1.35272358
#>  [8,] 0.2469038 0.5413271 0.13349130
#>  [9,] 1.0045660 1.6967409 2.03609988
#> [10,] 2.5913807 2.5913807 0.49080241
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.06505190 0.06505190 0.06505190
#>  [2,] 0.84996571 0.84996571 0.84996571
#>  [3,] 3.39237078 3.39237078 3.39237078
#>  [4,] 0.39846478 0.39846478 0.39846478
#>  [5,] 0.99519029 0.99519029 0.99519029
#>  [6,] 1.42054832 1.42054832 1.42054832
#>  [7,] 0.40984691 0.40984691 0.40984691
#>  [8,] 2.61046630 2.61046630 2.61046630
#>  [9,] 5.11212497 5.11212497 5.11212497
#> [10,] 0.03000942 0.03000942 0.03000942
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 0.34908445 0.38926099 2.2584466
#>  [2,] 0.81134305 0.21601978 0.6252653
#>  [3,] 0.33448720 0.38046931 0.2792629
#>  [4,] 0.86969675 0.78342629 0.6016754
#>  [5,] 0.71297583 1.00762482 0.3206249
#>  [6,] 0.05053209 4.05141768 0.0327297
#>  [7,] 1.03288587 0.62633259 1.0435013
#>  [8,] 0.59479799 0.46177905 1.3123929
#>  [9,] 0.16759146 0.01355083 1.8308044
#> [10,] 0.60252354 0.13812015 1.0737641

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.08487189 0.7647106 1.7043989
#>  [2,] 1.66152648 0.3504956 1.7107508
#>  [3,] 6.17969800 0.1010990 1.8282518
#>  [4,] 4.11501275 1.1107603 0.5976493
#>  [5,] 0.56380477 1.2244930 1.2244930
#>  [6,] 1.67551125 1.2290842 1.4460850
#>  [7,] 0.03903141 0.2249848 0.2249848
#>  [8,] 1.79401984 1.7940198 0.9635807
#>  [9,] 0.25106679 2.5874583 1.0855810
#> [10,] 0.92524324 4.4667584 0.6324657
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14197433 0.14197433 0.14197433
#>  [2,] 0.75045094 0.75045094 0.75045094
#>  [3,] 0.02382837 0.02382837 0.02382837
#>  [4,] 0.04887923 0.04887923 0.04887923
#>  [5,] 0.01948357 0.01948357 0.01948357
#>  [6,] 0.40538533 0.40538533 0.40538533
#>  [7,] 0.09092059 0.09092059 0.09092059
#>  [8,] 2.37263642 2.37263642 2.37263642
#>  [9,] 3.29197958 3.29197958 3.29197958
#> [10,] 2.47053317 2.47053317 2.47053317
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]      [,3]
#>  [1,] 2.17335808 0.4094465 3.7012840
#>  [2,] 1.42920474 1.7076854 2.1258611
#>  [3,] 4.43173997 1.1999467 0.8081377
#>  [4,] 0.25299137 1.9522189 0.5273320
#>  [5,] 1.42962330 1.2134269 0.3852644
#>  [6,] 0.95308653 2.0076690 2.1257439
#>  [7,] 0.64770936 0.2083105 0.5634004
#>  [8,] 0.04739604 0.2028500 0.1584779
#>  [9,] 0.25373345 0.7456542 2.6437312
#> [10,] 0.97490837 1.5705363 0.0403542

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.1349794 10.3308959 3.97682724
#>  [2,] 0.5621979  0.8261790 4.55686752
#>  [3,] 0.1710396  0.5971046 0.51659031
#>  [4,] 0.3451705  0.3451705 0.34517050
#>  [5,] 1.3028613  0.1551826 0.96348830
#>  [6,] 0.5413393  0.9482045 0.76560528
#>  [7,] 0.7553366  0.7553366 0.16054727
#>  [8,] 0.1558207  3.4640328 0.06007079
#>  [9,] 2.6256586  2.8791204 6.84183516
#> [10,] 0.5438619  0.5438619 0.54386193
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.80964498 1.80964498 1.80964498
#>  [2,] 0.38187758 0.38187758 0.38187758
#>  [3,] 0.08597891 0.08597891 0.08597891
#>  [4,] 3.40906071 3.40906071 3.40906071
#>  [5,] 1.56768804 1.56768804 1.56768804
#>  [6,] 0.08841230 0.08841230 0.08841230
#>  [7,] 3.08520961 3.08520961 3.08520961
#>  [8,] 0.33005915 0.33005915 0.33005915
#>  [9,] 1.73641212 1.73641212 1.73641212
#> [10,] 0.59894488 0.59894488 0.59894488
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>             [,1]      [,2]        [,3]
#>  [1,] 0.91877816 0.9964566 0.661522339
#>  [2,] 0.25737640 0.4416287 0.543128281
#>  [3,] 1.39625733 0.6331817 0.004229847
#>  [4,] 1.18191083 0.4632903 0.181129522
#>  [5,] 1.02135402 0.3424757 1.578931629
#>  [6,] 0.03663183 0.1666661 0.946381345
#>  [7,] 3.31531783 0.1036813 1.265409481
#>  [8,] 0.88286305 2.3548731 1.110411608
#>  [9,] 1.15223666 2.3228541 0.607264517
#> [10,] 0.66784657 0.5388932 1.014083084

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.01796272 0.08162666 0.08162666
#>  [2,] 1.98288899 5.76750670 0.37296820
#>  [3,] 0.84679986 2.79329227 0.43328533
#>  [4,] 1.33413998 1.33413998 0.11954967
#>  [5,] 0.49279789 0.13868653 1.53139580
#>  [6,] 1.19622895 2.09349028 2.75840943
#>  [7,] 1.56213272 3.29211108 0.83878648
#>  [8,] 0.50699356 0.50699356 0.50699356
#>  [9,] 1.67728702 0.71652712 1.67728702
#> [10,] 0.34166239 2.09876567 1.81857163
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.32586559 0.32586559 0.32586559
#>  [2,] 0.74765165 0.74765165 0.74765165
#>  [3,] 0.07952610 0.07952610 0.07952610
#>  [4,] 0.70103754 0.70103754 0.70103754
#>  [5,] 0.40232310 0.40232310 0.40232310
#>  [6,] 0.69684841 0.69684841 0.69684841
#>  [7,] 2.67646145 2.67646145 2.67646145
#>  [8,] 0.39129443 0.39129443 0.39129443
#>  [9,] 0.73981667 0.73981667 0.73981667
#> [10,] 0.07040459 0.07040459 0.07040459
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.01347218 1.20499550 2.36345120
#>  [2,] 0.02460431 0.06240596 0.10121114
#>  [3,] 0.23138337 0.41450393 2.10687525
#>  [4,] 0.84457218 1.91018027 1.23149250
#>  [5,] 0.66217889 1.49960248 2.10919469
#>  [6,] 0.25311593 0.16730979 0.11537481
#>  [7,] 0.52891246 0.62462011 1.18805511
#>  [8,] 0.17708903 1.35022387 0.48450840
#>  [9,] 0.33998810 1.31698561 0.06509349
#> [10,] 1.53655360 0.63175333 0.36222599

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.77710879 0.20813166 0.2081317
#>  [2,] 0.20689623 1.21820132 0.2068962
#>  [3,] 1.51275859 0.37913311 1.4421826
#>  [4,] 0.18957372 0.01358036 0.1895737
#>  [5,] 1.07315940 1.07315940 1.0731594
#>  [6,] 1.29395678 0.44587591 0.4458759
#>  [7,] 0.37990184 0.37990184 1.6258543
#>  [8,] 0.53838335 1.28169243 2.0624259
#>  [9,] 0.05659199 0.27897451 0.2627344
#> [10,] 0.02447416 0.38481040 0.3848104
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6606753 0.34576890 0.95399484
#>  [2,] 0.2053778 0.81050052 1.30200570
#>  [3,] 1.2298223 1.84397066 0.87901876
#>  [4,] 2.3115890 0.37526579 0.37526579
#>  [5,] 0.3518390 0.01199388 0.35183899
#>  [6,] 0.4090667 0.86554669 0.40906668
#>  [7,] 0.6044220 0.60442198 0.60442198
#>  [8,] 0.2214299 0.22142991 0.05642902
#>  [9,] 0.5586917 0.08876899 0.12983958
#> [10,] 1.1628568 0.20701509 1.92134284
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.39463314 0.39463314 0.3946331
#>  [2,] 1.46057617 1.36300750 0.9400150
#>  [3,] 0.77698237 0.07199081 0.6630363
#>  [4,] 0.75647643 0.93838040 0.9383804
#>  [5,] 0.34153002 1.22927934 0.3626290
#>  [6,] 3.37867388 0.37084070 0.2878365
#>  [7,] 0.25957395 1.29156151 0.2595740
#>  [8,] 0.30657498 0.14079204 0.6357117
#>  [9,] 0.05759635 0.05759635 0.8103815
#> [10,] 0.06381786 0.54261642 0.4918422
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.63387501 1.90027803 0.7468275
#>  [2,] 0.04335793 0.03490982 0.4245641
#>  [3,] 0.30032458 0.83693754 0.5042011
#>  [4,] 2.51775517 0.26872540 0.1558369
#>  [5,] 0.55344172 0.55344172 0.2151869
#>  [6,] 1.56902710 0.22060401 0.2206040
#>  [7,] 1.03332253 0.41510639 0.1717352
#>  [8,] 0.37068514 1.40945539 0.9018503
#>  [9,] 0.37872163 0.68261984 0.7531173
#> [10,] 0.08508705 0.52666220 0.5266622
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.34206840 0.3420684 0.23797413
#>  [2,] 0.02844117 1.7048825 0.02844117
#>  [3,] 1.24058073 1.1399473 1.24058073
#>  [4,] 1.62891182 1.3066951 0.55363911
#>  [5,] 0.29272633 1.8352597 1.04860555
#>  [6,] 0.17847131 0.1784713 0.02466653
#>  [7,] 1.48068941 0.5325503 0.53255027
#>  [8,] 0.50150974 0.5562464 0.55624637
#>  [9,] 0.03928330 0.0645490 0.03928330
#> [10,] 1.32181013 0.1254958 0.78513537

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.28825525 2.3542904 0.1518437
#>  [2,] 4.60732641 2.9145833 3.8567036
#>  [3,] 0.44049435 1.9506270 0.9819361
#>  [4,] 3.91017416 0.1635094 0.1227715
#>  [5,] 3.89152975 1.1390910 0.5125367
#>  [6,] 1.73123194 1.6632296 0.3554222
#>  [7,] 0.68329148 0.6346589 2.1346625
#>  [8,] 0.02649375 0.8497832 0.6691339
#>  [9,] 1.48788025 2.0628482 0.1012021
#> [10,] 1.76728833 1.3472197 0.7210997
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]      [,2]       [,3]
#>  [1,] 0.61074478 3.1311384 1.88416730
#>  [2,] 0.02375258 0.4532321 0.02375258
#>  [3,] 2.20917940 0.6752669 0.84304454
#>  [4,] 0.79452533 0.8217484 0.34765775
#>  [5,] 0.69098768 0.1201127 1.28473528
#>  [6,] 0.31067747 0.3106775 0.31067747
#>  [7,] 0.48554477 0.4855448 0.48554477
#>  [8,] 4.22953465 1.3018887 1.20307023
#>  [9,] 1.75879037 1.7587904 1.75879037
#> [10,] 0.02918209 1.6389486 1.63894860
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.10657861 0.6507262 0.6507262
#>  [2,] 6.52607637 2.6149531 0.8011283
#>  [3,] 0.30886635 0.3088663 0.3088663
#>  [4,] 7.30410172 5.3995820 3.0651459
#>  [5,] 0.10179424 0.1017942 0.1017942
#>  [6,] 0.48058914 0.0664689 0.4805891
#>  [7,] 0.47097794 3.6380073 0.3920471
#>  [8,] 0.00724879 5.2922142 1.2595950
#>  [9,] 0.60970382 1.5395245 1.5395245
#> [10,] 4.32394497 3.2644575 0.2335269
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.6846205 1.7960140 1.2210730
#>  [2,] 0.1453805 0.3590581 1.6790593
#>  [3,] 3.6451923 1.0807386 0.7224147
#>  [4,] 1.6073552 0.2882285 2.4756616
#>  [5,] 1.9622974 5.7053772 0.8548482
#>  [6,] 0.2784439 0.2784439 0.2784439
#>  [7,] 0.7618538 0.7618538 0.7618538
#>  [8,] 0.6810668 2.9424942 0.3011721
#>  [9,] 0.3233109 1.0502613 0.8881440
#> [10,] 3.8348719 0.5526536 1.1954384
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.8232877 0.64892671 1.8232877
#>  [2,] 0.4329002 0.43290023 0.2512267
#>  [3,] 1.6824824 0.07929821 0.3091153
#>  [4,] 1.0649253 1.06492533 1.0649253
#>  [5,] 3.4181528 1.87912334 3.0185672
#>  [6,] 0.1225695 0.12256949 0.1225695
#>  [7,] 3.1584439 0.44108349 1.3807637
#>  [8,] 0.2842450 0.28424505 0.2842450
#>  [9,] 2.3942915 0.33297904 0.1373796
#> [10,] 0.4411938 2.18170848 1.0116775

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.70714844 0.34688692 0.39437494
#>  [2,] 0.08286997 0.08286997 0.08286997
#>  [3,] 0.03085616 0.15960953 0.15960953
#>  [4,] 0.43303752 0.43303752 0.43303752
#>  [5,] 0.01528346 0.01528346 0.01528346
#>  [6,] 0.12557180 0.12557180 0.12557180
#>  [7,] 0.31819104 0.31819104 0.30808777
#>  [8,] 0.10445016 0.10445016 0.10445016
#>  [9,] 0.05252964 0.22718145 0.20608052
#> [10,] 0.01323922 0.11983933 0.01323922
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2993232 0.29932324 0.2993232
#>  [2,] 0.5142287 0.51422869 0.5142287
#>  [3,] 0.2373773 0.24055416 0.2373773
#>  [4,] 0.1626285 0.16262852 0.1626285
#>  [5,] 0.7725381 0.02911093 0.5209989
#>  [6,] 0.5316427 0.31953147 0.5316427
#>  [7,] 0.1274624 0.13212764 0.1321276
#>  [8,] 1.1132336 2.60925369 1.8126132
#>  [9,] 0.2976630 0.50439594 0.5043959
#> [10,] 0.4426477 0.10111930 0.1011193
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.52404946 0.42915600 0.52404946
#>  [2,] 0.32422748 0.32422748 0.32422748
#>  [3,] 1.12900525 0.56885777 0.56885777
#>  [4,] 1.86088027 1.86088027 0.90468829
#>  [5,] 0.02041718 0.32771527 1.09720174
#>  [6,] 0.24452270 0.24452270 0.24452270
#>  [7,] 0.03048069 0.03048069 0.03048069
#>  [8,] 0.36326535 0.79541256 0.01603933
#>  [9,] 0.51342018 0.62059654 0.62059654
#> [10,] 0.65311072 0.50947847 0.65311072
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]        [,2]      [,3]
#>  [1,] 0.28005033 1.254769590 1.2547696
#>  [2,] 0.20538999 0.205389990 0.1181587
#>  [3,] 0.06802191 0.647964020 0.5000244
#>  [4,] 0.41922816 0.006348866 0.3127210
#>  [5,] 0.14454635 0.144546346 0.1445463
#>  [6,] 0.11792711 0.558917200 0.8997272
#>  [7,] 0.29150456 0.241871900 0.2915046
#>  [8,] 0.10990979 0.109909787 0.1099098
#>  [9,] 0.42857565 0.450756378 0.3967205
#> [10,] 0.45872648 0.894363632 0.3891823
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02447916 0.02447916 0.02447916
#>  [2,] 0.55318331 0.26859777 0.26859777
#>  [3,] 0.71100722 0.82993597 0.82993597
#>  [4,] 0.60116506 0.11239847 0.49947385
#>  [5,] 0.02697367 2.74173339 0.43142756
#>  [6,] 0.01479966 0.01479966 0.01479966
#>  [7,] 0.27784952 0.27784952 0.27784952
#>  [8,] 1.25945632 1.93882446 1.93882446
#>  [9,] 0.53698097 0.39668891 0.53698097
#> [10,] 0.59952294 0.59952294 0.09581474

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.95932162 1.47513578 0.18380412
#>  [2,] 0.08439807 0.28683488 0.77833632
#>  [3,] 0.06682155 0.31019454 0.06682155
#>  [4,] 0.40566337 0.18793474 0.40566337
#>  [5,] 1.16832469 1.26517531 0.07791612
#>  [6,] 0.31920610 0.31920610 0.31920610
#>  [7,] 0.56047484 0.40254345 0.40254345
#>  [8,] 0.28701681 0.25663584 0.34043407
#>  [9,] 0.08186056 0.08186056 0.08186056
#> [10,] 0.89039750 0.04343738 0.78269135
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.04110159 0.86512323 0.17845612
#>  [2,] 0.01761578 0.01761578 0.01761578
#>  [3,] 0.05924532 0.16926296 0.07601065
#>  [4,] 0.04975617 0.30554368 0.04975617
#>  [5,] 0.10040161 0.10040161 0.10040161
#>  [6,] 0.23857366 0.46859801 0.46859801
#>  [7,] 0.42999071 0.48419154 0.48419154
#>  [8,] 0.49321429 0.49321429 0.58241711
#>  [9,] 0.18963627 0.06456992 0.12919380
#> [10,] 0.26652355 1.42206033 0.31452006
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.14551047 0.14551047 0.002659496
#>  [2,] 0.24071132 0.24071132 0.067717174
#>  [3,] 0.20955831 0.20955831 0.209558312
#>  [4,] 0.21517516 0.40378810 1.147666489
#>  [5,] 0.28455659 0.68245015 0.003684854
#>  [6,] 0.23310836 0.23310836 0.125019672
#>  [7,] 0.24446397 0.24446397 0.244463973
#>  [8,] 0.05336125 0.05336125 0.195019560
#>  [9,] 0.18520718 0.10911551 0.109115508
#> [10,] 0.19497513 0.19497513 0.012046866
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.27806921 0.36085272 0.05805145
#>  [2,] 0.69187017 0.68743827 0.68743827
#>  [3,] 0.53539816 1.26722029 0.66241196
#>  [4,] 0.62706568 0.45770361 0.09346677
#>  [5,] 1.23087288 0.67829013 0.37124115
#>  [6,] 0.06728321 0.06728321 0.65308458
#>  [7,] 0.15384066 0.14510907 0.46613501
#>  [8,] 0.39754598 0.61684866 0.19518892
#>  [9,] 0.09542506 0.68339796 0.75923608
#> [10,] 0.15786389 0.55433550 0.83517861

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.82024293 0.36148790 0.3109537
#>  [2,] 0.53938374 0.42129919 0.2440769
#>  [3,] 0.48643277 0.57959315 0.4864328
#>  [4,] 0.01063303 0.28687844 0.2868784
#>  [5,] 0.29358095 0.38861143 0.3886114
#>  [6,] 0.24879436 0.82160328 0.3333376
#>  [7,] 0.15848953 0.74973108 1.2942454
#>  [8,] 0.20301930 0.32801773 0.1427133
#>  [9,] 0.58765259 0.25445176 0.6178771
#> [10,] 0.06661705 0.06661705 0.9563110
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1616118 0.03806169 0.1616118
#>  [2,] 0.2951248 0.17081315 1.1250941
#>  [3,] 0.5750019 0.76716954 0.7156244
#>  [4,] 0.2970491 0.29704911 0.2970491
#>  [5,] 0.1344983 0.27604098 0.1344983
#>  [6,] 0.1870693 0.18706934 0.1870693
#>  [7,] 0.3245617 0.32456169 0.3245617
#>  [8,] 1.4971843 0.01086961 2.1120201
#>  [9,] 0.3166568 0.13754008 0.3166568
#> [10,] 0.3031420 0.30314196 0.3031420
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.10225969 0.10225969 0.1022597
#>  [2,] 0.39005247 0.36786138 0.3900525
#>  [3,] 0.47443914 0.47443914 0.4744391
#>  [4,] 0.45142179 0.45142179 0.2765307
#>  [5,] 0.46621080 0.78526441 0.7320773
#>  [6,] 0.27649690 0.25675913 0.6826447
#>  [7,] 0.02598334 0.06985742 0.3493583
#>  [8,] 1.13246603 0.44498555 0.1048934
#>  [9,] 0.11788808 0.46901371 0.1178881
#> [10,] 0.24878681 0.46906373 0.3432936
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09592167 0.85866483 0.26093525
#>  [2,] 0.09463949 0.09463949 0.09463949
#>  [3,] 0.09606956 0.34376673 0.11782054
#>  [4,] 1.59431681 1.93875394 0.95863501
#>  [5,] 1.31381839 1.67709498 1.31769243
#>  [6,] 0.19454957 0.28755799 0.28755799
#>  [7,] 0.94000394 0.26000699 0.15593851
#>  [8,] 0.24602443 0.24602443 1.07043447
#>  [9,] 0.17059326 0.72261879 0.72261879
#> [10,] 0.20676496 0.20676496 0.40044515

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>             [,1]       [,2]        [,3]
#>  [1,] 0.71017115 0.35358965 0.710171149
#>  [2,] 0.27088026 0.27088026 0.001467723
#>  [3,] 0.63133166 0.62122862 0.631331658
#>  [4,] 0.94257558 0.17087438 0.116160775
#>  [5,] 0.49789498 0.15682434 0.287815704
#>  [6,] 0.07184981 0.07184981 0.071849813
#>  [7,] 0.51316734 0.03724032 0.028668023
#>  [8,] 0.34183522 0.34183522 0.341835220
#>  [9,] 0.17238195 0.17238195 0.172381950
#> [10,] 0.14731363 0.14731363 0.147313627
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]       [,2]      [,3]
#>  [1,] 1.71229773 0.37376108 3.4332078
#>  [2,] 0.29481109 0.40743246 0.3611644
#>  [3,] 0.66523081 0.19551231 0.2284131
#>  [4,] 0.05199641 0.06468073 0.3148731
#>  [5,] 0.26342498 0.26342498 0.7137474
#>  [6,] 0.85444286 0.57103054 0.1166426
#>  [7,] 0.36696942 0.36696942 0.3669694
#>  [8,] 0.23203312 0.04981486 0.2320331
#>  [9,] 0.11925802 1.10303431 0.3212343
#> [10,] 0.62713447 0.22831356 0.1553014
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3826979 0.7720410 0.7720410
#>  [2,] 0.5744683 0.8533284 0.1902447
#>  [3,] 0.1174294 0.2258105 0.2258105
#>  [4,] 0.2028286 0.5806209 0.7469163
#>  [5,] 0.1203081 0.1203081 0.1203081
#>  [6,] 0.4313000 0.4313000 0.4313000
#>  [7,] 0.1578487 0.3658665 0.6084414
#>  [8,] 0.2048860 0.2048860 0.2048860
#>  [9,] 0.6559086 0.6844226 0.6844226
#> [10,] 0.8191267 0.3355348 0.8699710
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.204376600 0.502630547 0.502630547
#>  [2,] 0.001122763 0.001122763 0.001122763
#>  [3,] 0.166572765 0.119279937 0.709181864
#>  [4,] 0.565066498 0.149278826 0.526483703
#>  [5,] 0.113400540 0.113400540 0.113400540
#>  [6,] 0.294237911 0.294237911 0.201883697
#>  [7,] 0.106334667 0.200370589 0.298565905
#>  [8,] 0.038737129 0.038737129 0.038737129
#>  [9,] 0.438482792 0.078595056 0.270439119
#> [10,] 0.980135411 0.333086022 0.980135411