Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi: 10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]       [,2]       [,3]
#>  [1,] 1.91593186 0.85825580 0.26291965
#>  [2,] 3.09639137 0.13841630 3.09639137
#>  [3,] 0.99113869 0.99113869 0.99113869
#>  [4,] 0.56637221 1.68217193 1.23337375
#>  [5,] 0.24160845 0.96227766 0.94628985
#>  [6,] 8.12867049 1.43113063 0.47075127
#>  [7,] 0.22247369 0.22247369 0.22247369
#>  [8,] 0.72787484 0.54166555 0.46932990
#>  [9,] 4.57416745 1.40268118 1.13034876
#> [10,] 0.04431013 0.04431013 0.04431013
rpextmo(10, 3, a = 1)      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.62781864 1.62781864 1.62781864
#>  [2,] 0.07152678 0.07152678 0.07152678
#>  [3,] 0.55390054 0.55390054 0.55390054
#>  [4,] 0.64768767 0.64768767 0.64768767
#>  [5,] 0.39645122 0.39645122 0.39645122
#>  [6,] 0.96004298 0.96004298 0.96004298
#>  [7,] 0.71199382 0.71199382 0.71199382
#>  [8,] 0.09538560 0.09538560 0.09538560
#>  [9,] 0.87071843 0.87071843 0.87071843
#> [10,] 0.09704352 0.09704352 0.09704352
rpextmo(10, 3, b = 1)      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.6274679 1.3498578 0.32228176
#>  [2,] 0.3506600 0.8931130 0.26469734
#>  [3,] 0.1753614 1.9621867 1.01372004
#>  [4,] 1.8548152 2.6685865 0.71617676
#>  [5,] 1.2265249 0.5879140 0.80738886
#>  [6,] 0.4559077 0.9270563 2.03338280
#>  [7,] 3.1909552 0.5053513 0.40808868
#>  [8,] 0.3662007 1.9259042 0.08191946
#>  [9,] 0.2505683 1.9815261 0.49248538
#> [10,] 0.7632544 0.2846699 0.95854719

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.1251782 2.30067011 0.1856896
#>  [2,] 2.1432996 2.29699185 0.1812302
#>  [3,] 0.3719909 0.02844419 5.3726399
#>  [4,] 2.0385757 0.08995068 0.4441913
#>  [5,] 0.5819011 0.58190106 0.5491798
#>  [6,] 1.8685348 0.69641630 0.5177420
#>  [7,] 1.4840157 1.70131958 0.8879171
#>  [8,] 1.8589526 2.66773055 1.4011922
#>  [9,] 0.2371530 0.80997718 1.1271554
#> [10,] 1.2495663 1.24956626 0.1574295
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.25796246 0.25796246 0.25796246
#>  [2,] 0.72945539 0.72945539 0.72945539
#>  [3,] 0.13446752 0.13446752 0.13446752
#>  [4,] 0.79429722 0.79429722 0.79429722
#>  [5,] 0.00246227 0.00246227 0.00246227
#>  [6,] 0.36023514 0.36023514 0.36023514
#>  [7,] 0.25386919 0.25386919 0.25386919
#>  [8,] 0.09021082 0.09021082 0.09021082
#>  [9,] 1.17798124 1.17798124 1.17798124
#> [10,] 0.13786674 0.13786674 0.13786674
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 1.36437841 0.07105546 0.30825881
#>  [2,] 0.04750656 1.05966691 0.24740483
#>  [3,] 0.13999036 1.08677906 0.10618515
#>  [4,] 0.91285812 0.36499559 0.35946819
#>  [5,] 1.03934711 0.46601387 0.04767924
#>  [6,] 1.60127869 1.42910819 0.63342353
#>  [7,] 1.48117674 1.37503901 0.16099165
#>  [8,] 1.84233077 2.76027928 2.04011423
#>  [9,] 0.41849113 0.03412489 0.53466352
#> [10,] 1.73324160 0.69811185 0.18246463

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.42052332 6.62559982 2.59586318
#>  [2,] 1.69035496 1.82744394 4.34423639
#>  [3,] 0.07771421 0.07771421 0.07771421
#>  [4,] 1.88221856 1.28604489 0.91982074
#>  [5,] 0.40307070 0.40307070 0.40307070
#>  [6,] 2.02454068 2.02454068 2.02454068
#>  [7,] 2.59617254 1.26769830 1.50338649
#>  [8,] 0.45563226 0.45563226 0.45563226
#>  [9,] 0.38696669 0.05958922 0.20534594
#> [10,] 0.99028704 0.80797359 0.43842889
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4213966 0.4213966 0.4213966
#>  [2,] 0.4197234 0.4197234 0.4197234
#>  [3,] 0.4133909 0.4133909 0.4133909
#>  [4,] 1.3892065 1.3892065 1.3892065
#>  [5,] 1.8411553 1.8411553 1.8411553
#>  [6,] 0.8988425 0.8988425 0.8988425
#>  [7,] 0.1976729 0.1976729 0.1976729
#>  [8,] 0.3992708 0.3992708 0.3992708
#>  [9,] 1.3262013 1.3262013 1.3262013
#> [10,] 0.5703437 0.5703437 0.5703437
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8135883 1.69433234 1.4136568
#>  [2,] 0.3173165 0.35999014 0.1901662
#>  [3,] 0.4442261 1.93327542 0.8498059
#>  [4,] 1.0571463 0.35889972 0.5401353
#>  [5,] 0.1713098 0.80696334 4.3004414
#>  [6,] 0.6970863 1.07658397 0.7368997
#>  [7,] 0.1541943 0.31184143 0.4796142
#>  [8,] 0.6841357 0.42694767 0.6443594
#>  [9,] 1.3222284 2.66567868 1.1906996
#> [10,] 1.3008874 0.05372877 0.7834726

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>              [,1]      [,2]      [,3]
#>  [1,] 3.334312650 2.6534232 0.4396127
#>  [2,] 1.361497159 1.2532891 0.2117941
#>  [3,] 0.007954434 0.4914986 0.7379871
#>  [4,] 2.049151433 2.0491514 0.6597728
#>  [5,] 1.471841561 1.4094317 0.1156602
#>  [6,] 0.588384599 3.1858295 3.1858295
#>  [7,] 0.456316057 0.2125002 0.5380203
#>  [8,] 0.114945294 0.1149453 0.1149453
#>  [9,] 0.379849207 0.8385602 1.1577522
#> [10,] 1.730317993 0.5593198 1.0725620
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.47557289 0.47557289 0.47557289
#>  [2,] 0.53819987 0.53819987 0.53819987
#>  [3,] 0.02473702 0.02473702 0.02473702
#>  [4,] 0.54861198 0.54861198 0.54861198
#>  [5,] 0.61589279 0.61589279 0.61589279
#>  [6,] 0.50804845 0.50804845 0.50804845
#>  [7,] 0.75067466 0.75067466 0.75067466
#>  [8,] 0.83790707 0.83790707 0.83790707
#>  [9,] 0.60469119 0.60469119 0.60469119
#> [10,] 2.22852741 2.22852741 2.22852741
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>             [,1]       [,2]        [,3]
#>  [1,] 0.66228495 0.27208402 2.889150451
#>  [2,] 0.18947493 0.08167095 0.303150762
#>  [3,] 1.31423270 1.07838098 0.177631707
#>  [4,] 0.37367838 1.15819308 0.753626643
#>  [5,] 0.70411417 1.62836247 0.003574234
#>  [6,] 1.77788337 0.14053155 0.067700710
#>  [7,] 2.34465483 0.01714873 0.089038940
#>  [8,] 0.06760596 0.32669708 0.033202886
#>  [9,] 0.68632682 0.14289034 0.437750672
#> [10,] 1.49005093 1.84721656 0.658793248

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.84531619 1.44333323 0.47694647
#>  [2,] 0.25671299 1.66468952 4.13085561
#>  [3,] 1.45007434 0.42468655 0.55740782
#>  [4,] 2.25066724 0.44986407 2.78767423
#>  [5,] 0.82318667 0.82318667 0.57659516
#>  [6,] 3.08995659 0.09301540 3.08995659
#>  [7,] 0.63116974 0.63116974 0.63116974
#>  [8,] 0.60462082 0.15515792 0.60462082
#>  [9,] 0.65003077 0.04972503 0.23826898
#> [10,] 0.01200967 0.01200967 0.01200967
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.008535106 0.008535106 0.008535106
#>  [2,] 0.672120946 0.672120946 0.672120946
#>  [3,] 0.155028180 0.155028180 0.155028180
#>  [4,] 0.280112160 0.280112160 0.280112160
#>  [5,] 2.204074622 2.204074622 2.204074622
#>  [6,] 0.423806835 0.423806835 0.423806835
#>  [7,] 6.119871482 6.119871482 6.119871482
#>  [8,] 0.928237011 0.928237011 0.928237011
#>  [9,] 0.383325805 0.383325805 0.383325805
#> [10,] 1.125886331 1.125886331 1.125886331
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>             [,1]       [,2]       [,3]
#>  [1,] 0.50707733 0.03922531 0.28288928
#>  [2,] 0.08390603 0.06442474 1.34818609
#>  [3,] 0.09664214 2.19201359 1.61268055
#>  [4,] 0.51465323 2.38637122 0.33210256
#>  [5,] 1.82264947 0.17900043 0.02513937
#>  [6,] 0.91304739 0.58887545 3.42687181
#>  [7,] 0.10951539 0.25952450 0.90730167
#>  [8,] 3.11445044 0.71704171 0.41965506
#>  [9,] 1.36811316 2.00382006 2.76290921
#> [10,] 0.54296788 1.75705071 2.51616941

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4864836 0.87756929 0.1176424
#>  [2,] 1.5953507 0.57988324 1.5953507
#>  [3,] 0.1428852 0.06742904 1.0489672
#>  [4,] 1.5684832 0.85221916 1.5177967
#>  [5,] 0.3490286 0.15272388 0.3490286
#>  [6,] 1.7831248 0.14359804 1.4678822
#>  [7,] 0.1638671 0.16386707 0.1638671
#>  [8,] 0.1888011 0.53554660 1.4559387
#>  [9,] 0.3736927 1.26392319 1.6986837
#> [10,] 2.3335006 0.09964832 5.2458897
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 1.027635575 1.027635575 1.027635575
#>  [2,] 0.570998986 0.570998986 0.570998986
#>  [3,] 1.129636834 1.129636834 1.129636834
#>  [4,] 1.097236925 1.097236925 1.097236925
#>  [5,] 0.138305772 0.138305772 0.138305772
#>  [6,] 0.505610311 0.505610311 0.505610311
#>  [7,] 0.102479927 0.102479927 0.102479927
#>  [8,] 0.007468547 0.007468547 0.007468547
#>  [9,] 1.747590208 1.747590208 1.747590208
#> [10,] 1.076194233 1.076194233 1.076194233
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.5815841 0.03818214 0.74344374
#>  [2,] 0.5252012 0.80187893 0.64287735
#>  [3,] 3.1186195 0.45437953 0.26741808
#>  [4,] 1.6223404 1.15883786 3.09249162
#>  [5,] 2.7642248 0.09061598 0.96658280
#>  [6,] 0.4574772 0.18193466 0.08604742
#>  [7,] 0.6997482 4.22168418 2.39296654
#>  [8,] 0.7957629 1.87287704 1.19378661
#>  [9,] 0.1865442 2.02547507 0.41057365
#> [10,] 1.1788774 1.27056081 0.39612905

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 0.7596729 1.6604826 1.2546392
#>  [2,] 0.2016603 0.1867843 1.1168847
#>  [3,] 3.0039431 1.7519825 4.9773626
#>  [4,] 0.4833654 1.1809119 0.4776747
#>  [5,] 1.9456253 0.5084773 2.1977436
#>  [6,] 0.9479700 1.7770480 1.3830414
#>  [7,] 1.9485118 0.1290078 1.9970230
#>  [8,] 1.3953730 1.0282934 4.7781528
#>  [9,] 0.1293495 0.8484838 0.9461581
#> [10,] 0.1149468 0.4426974 3.0402419
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.29329648 1.29329648 1.29329648
#>  [2,] 0.72437709 0.72437709 0.72437709
#>  [3,] 0.16939254 0.16939254 0.16939254
#>  [4,] 0.39979922 0.39979922 0.39979922
#>  [5,] 0.05524111 0.05524111 0.05524111
#>  [6,] 0.22968262 0.22968262 0.22968262
#>  [7,] 2.50346911 2.50346911 2.50346911
#>  [8,] 0.62972006 0.62972006 0.62972006
#>  [9,] 0.43035799 0.43035799 0.43035799
#> [10,] 0.68842061 0.68842061 0.68842061
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>            [,1]      [,2]       [,3]
#>  [1,] 0.3653962 0.9742367 1.00732817
#>  [2,] 1.1890513 1.2655537 1.35559821
#>  [3,] 0.3935337 0.2374292 0.17133264
#>  [4,] 1.8342147 0.3771976 4.61611076
#>  [5,] 0.3268436 0.3584201 2.33192264
#>  [6,] 1.6981059 0.7613116 1.70280588
#>  [7,] 0.4365276 0.7695373 1.87401154
#>  [8,] 1.3767309 0.8665472 1.12678564
#>  [9,] 0.3957154 1.7760116 0.04144619
#> [10,] 0.9909807 0.2315169 0.12066860

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.2949503 0.29495026 0.2949503
#>  [2,] 1.1602668 2.01137164 1.0463693
#>  [3,] 0.1337826 0.13378257 0.1337826
#>  [4,] 2.9802461 1.34159955 0.3055823
#>  [5,] 0.5326351 0.53263510 0.5326351
#>  [6,] 0.8806371 0.73233805 0.1225612
#>  [7,] 0.6277749 0.06481997 1.0495862
#>  [8,] 1.5586210 1.55862099 1.5586210
#>  [9,] 2.8673067 2.86730665 0.6391273
#> [10,] 1.8166855 1.36484777 1.8778982
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 2.07819550 2.07819550 2.07819550
#>  [2,] 0.63734768 0.63734768 0.63734768
#>  [3,] 0.60013149 0.60013149 0.60013149
#>  [4,] 0.14580200 0.14580200 0.14580200
#>  [5,] 3.39318943 3.39318943 3.39318943
#>  [6,] 0.07503312 0.07503312 0.07503312
#>  [7,] 0.30912417 0.30912417 0.30912417
#>  [8,] 1.18431328 1.18431328 1.18431328
#>  [9,] 1.26637291 1.26637291 1.26637291
#> [10,] 2.08945624 2.08945624 2.08945624
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.01252789 1.3158075 1.50680484
#>  [2,] 0.38189922 0.6652489 0.56499052
#>  [3,] 2.44932393 5.3814448 0.84764744
#>  [4,] 0.23972333 1.6137622 0.87506925
#>  [5,] 0.56621288 0.4450106 1.31728774
#>  [6,] 2.00754350 1.2530065 0.41070551
#>  [7,] 0.91881357 0.2514151 2.63203611
#>  [8,] 1.87998238 0.8672072 0.07367212
#>  [9,] 0.13847213 0.1980125 1.02337302
#> [10,] 0.24349732 2.1576744 0.20777939

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.03728865 1.39380391 0.21712345
#>  [2,] 3.08081879 2.02701732 2.93180706
#>  [3,] 0.47232507 2.19835820 3.28186233
#>  [4,] 0.16875842 0.01909992 0.59158770
#>  [5,] 0.07378714 0.07378714 0.07378714
#>  [6,] 3.39216579 0.78747350 0.56800098
#>  [7,] 1.43531955 2.30766189 0.59031288
#>  [8,] 0.05870960 2.92353248 1.33894057
#>  [9,] 0.36088617 5.62371768 0.50859431
#> [10,] 0.48610177 1.96281608 0.67825432
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.2637036 0.2637036 0.2637036
#>  [2,] 0.1346382 0.1346382 0.1346382
#>  [3,] 0.3530498 0.3530498 0.3530498
#>  [4,] 4.4074811 4.4074811 4.4074811
#>  [5,] 0.1171359 0.1171359 0.1171359
#>  [6,] 1.9405598 1.9405598 1.9405598
#>  [7,] 1.0912777 1.0912777 1.0912777
#>  [8,] 3.5855796 3.5855796 3.5855796
#>  [9,] 4.7507045 4.7507045 4.7507045
#> [10,] 0.8855085 0.8855085 0.8855085
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 1.7882178 0.16466338 0.8675031
#>  [2,] 0.8417991 0.35964712 0.6117951
#>  [3,] 3.8314812 0.20629846 0.1895660
#>  [4,] 0.1848443 0.49742742 0.2539602
#>  [5,] 2.6281028 1.20920488 1.0242044
#>  [6,] 1.3915256 0.03961879 0.1581092
#>  [7,] 2.2525682 1.13954006 1.6938597
#>  [8,] 0.2391495 0.35664207 1.6048975
#>  [9,] 0.8078184 0.73044642 1.7370530
#> [10,] 0.0693854 1.50239493 2.8964747

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>              [,1]       [,2]      [,3]
#>  [1,] 2.104440194 6.84909405 0.6996339
#>  [2,] 2.313359062 5.14328983 0.3934731
#>  [3,] 0.033341026 1.33395353 1.3339535
#>  [4,] 0.133344708 0.06891582 0.1333447
#>  [5,] 2.726856898 1.54416717 1.2203165
#>  [6,] 0.003762102 1.25518203 3.5553719
#>  [7,] 0.049921080 0.82356806 3.6775220
#>  [8,] 0.690721380 3.93131576 1.1923872
#>  [9,] 0.842610791 1.18555015 4.1167236
#> [10,] 1.135234746 1.89689216 3.3785636
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 1.57773664 1.57773664 1.57773664
#>  [2,] 0.42161856 0.42161856 0.42161856
#>  [3,] 0.70535191 0.70535191 0.70535191
#>  [4,] 1.26445774 1.26445774 1.26445774
#>  [5,] 0.08569749 0.08569749 0.08569749
#>  [6,] 0.51922345 0.51922345 0.51922345
#>  [7,] 0.25331947 0.25331947 0.25331947
#>  [8,] 1.06534141 1.06534141 1.06534141
#>  [9,] 0.11147087 0.11147087 0.11147087
#> [10,] 0.69620603 0.69620603 0.69620603
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>             [,1]        [,2]      [,3]
#>  [1,] 0.38349573 0.224919980 2.6295222
#>  [2,] 0.30819164 2.041023574 5.4470386
#>  [3,] 0.36880639 3.165088302 0.1569386
#>  [4,] 0.36459448 0.052580783 0.1826914
#>  [5,] 1.78878727 0.004020332 0.8268244
#>  [6,] 0.69624346 0.192175766 0.8457308
#>  [7,] 0.02530988 0.342991406 0.2137045
#>  [8,] 0.07304683 0.183439193 0.1572857
#>  [9,] 0.20118676 0.738391053 5.7351076
#> [10,] 0.05635129 0.175215786 0.4355154

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>            [,1]      [,2]       [,3]
#>  [1,] 0.8641061 0.3346067 0.19793655
#>  [2,] 1.2620270 0.2922881 0.30167710
#>  [3,] 0.0868256 0.5099159 1.59011473
#>  [4,] 0.0747667 0.3559405 1.83758562
#>  [5,] 0.0140173 0.0140173 0.01401730
#>  [6,] 0.7379053 1.1701723 0.36686855
#>  [7,] 0.6113184 0.6113184 0.61131836
#>  [8,] 0.5598957 0.5598957 0.02906777
#>  [9,] 1.9885953 1.7984307 1.42468939
#> [10,] 1.8179894 0.8724935 2.38364336
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1963534 0.19635335 0.19635335
#>  [2,] 2.2546819 0.03289799 0.76585598
#>  [3,] 0.2224367 0.24202540 0.32847357
#>  [4,] 0.5722038 0.23108044 0.65533962
#>  [5,] 0.3160607 0.31606066 0.20418604
#>  [6,] 1.0936249 1.04018843 1.04018843
#>  [7,] 0.2095036 0.01738661 0.01738661
#>  [8,] 0.3133420 1.57259557 0.31334201
#>  [9,] 0.1153751 0.11537512 0.79570568
#> [10,] 0.1449970 0.58369751 1.09030376
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>              [,1]      [,2]       [,3]
#>  [1,] 1.387197767 0.2754574 0.27545739
#>  [2,] 3.140249806 1.2709006 1.27090060
#>  [3,] 0.001275197 1.3503405 0.10809653
#>  [4,] 2.818403763 2.9046281 1.77172405
#>  [5,] 0.464699861 1.9478928 0.16894848
#>  [6,] 0.178124486 0.3305424 0.06545341
#>  [7,] 0.076277277 0.8403423 0.07627728
#>  [8,] 0.250694852 0.3250547 0.17478051
#>  [9,] 0.128512190 3.1704645 0.89672162
#> [10,] 0.904079171 0.2431108 1.92773412
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.82951580 0.89156704 0.59433081
#>  [2,] 0.09682217 0.02603849 0.09682217
#>  [3,] 0.24906089 0.39416408 0.12714950
#>  [4,] 0.67384117 1.63485264 0.40109937
#>  [5,] 0.04315594 0.04315594 0.31167947
#>  [6,] 0.72803496 0.65535875 0.49929857
#>  [7,] 0.43613640 1.49945989 0.90445623
#>  [8,] 1.40489042 0.09508974 0.09508974
#>  [9,] 0.20144584 0.20144584 0.36947217
#> [10,] 0.13605280 1.51596927 0.09389770
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1162290 1.85820496 0.5553188
#>  [2,] 0.0688225 0.06882250 0.0688225
#>  [3,] 0.4154278 0.22119853 0.2211985
#>  [4,] 1.0981785 0.03578796 1.7399103
#>  [5,] 1.2623458 0.16865838 0.1686584
#>  [6,] 0.3958806 0.46977442 0.4697744
#>  [7,] 0.4932839 0.49328387 0.3360430
#>  [8,] 1.8050970 1.80509703 0.4353005
#>  [9,] 0.5287720 0.67673218 0.5287720
#> [10,] 1.3199474 0.97383661 1.1878502

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.6114787 0.2736625 1.6273782
#>  [2,] 2.3341916 1.2341317 4.8506960
#>  [3,] 1.1906023 1.7002220 0.5211003
#>  [4,] 0.7094937 0.2729431 0.2565380
#>  [5,] 1.8619583 1.8321648 2.6529136
#>  [6,] 0.5358866 0.5130787 0.5358866
#>  [7,] 0.1109028 0.1109028 0.1109028
#>  [8,] 4.4690193 4.2231963 5.5403794
#>  [9,] 1.5540859 1.6025307 0.4284238
#> [10,] 0.3473029 0.3473029 0.1864495
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7418007 2.3595946 5.9144573
#>  [2,] 0.9955402 0.9955402 0.9955402
#>  [3,] 0.5882223 0.8677105 2.1519623
#>  [4,] 0.4596991 0.2297926 1.3111971
#>  [5,] 3.7188143 1.9247182 1.9896277
#>  [6,] 0.7196050 5.4065615 1.0583413
#>  [7,] 3.0470754 1.8297303 0.4067444
#>  [8,] 0.6175329 2.4008848 3.4105691
#>  [9,] 1.2708907 1.6036677 0.9276002
#> [10,] 2.4474162 3.5873307 1.3371229
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 7.44427938 3.67213016 2.61900744
#>  [2,] 1.46460665 0.34425530 2.88815885
#>  [3,] 1.64364670 0.84354439 1.86559881
#>  [4,] 1.01691328 1.01691328 1.01691328
#>  [5,] 0.76314715 0.04788869 3.13769699
#>  [6,] 4.35556437 0.46032176 6.10100028
#>  [7,] 0.23113583 0.24841801 0.24841801
#>  [8,] 0.41194106 0.41194106 0.08710454
#>  [9,] 0.08840776 0.08840776 0.08840776
#> [10,] 0.11229274 1.34167399 1.46032680
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23903424 0.07378321 0.30748538
#>  [2,] 1.66842440 1.04624088 1.66842440
#>  [3,] 0.13535212 1.81062061 0.23662018
#>  [4,] 0.04335071 0.04335071 0.04335071
#>  [5,] 0.08364348 1.18259981 1.96612615
#>  [6,] 0.13848592 0.56286726 1.63063653
#>  [7,] 0.16865510 2.91312782 0.03297284
#>  [8,] 1.22285889 2.48792046 4.04295616
#>  [9,] 0.42552584 3.54230619 2.47690766
#> [10,] 3.88950785 6.18829343 7.02685157
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>            [,1]      [,2]       [,3]
#>  [1,] 1.2106141 2.3121324 0.34226097
#>  [2,] 0.3116085 1.1766095 0.31160850
#>  [3,] 0.8193551 0.4372224 0.09618769
#>  [4,] 0.8313725 0.5340432 0.19991032
#>  [5,] 1.2275084 1.2275084 1.22750845
#>  [6,] 1.6811567 0.5070697 1.52516924
#>  [7,] 1.8299964 1.0613614 0.13100188
#>  [8,] 0.4754582 2.5438423 2.34659977
#>  [9,] 3.8845351 2.1681614 1.29965847
#> [10,] 0.7863063 0.3390416 1.19272477

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.13122998 0.13122998 0.13122998
#>  [2,] 1.13893736 0.23919749 0.23919749
#>  [3,] 0.59926060 0.13523259 0.59926060
#>  [4,] 0.17040938 0.23593771 0.23593771
#>  [5,] 0.07471043 1.33902613 0.22157704
#>  [6,] 1.21930981 0.24774148 1.21930981
#>  [7,] 0.09687582 0.09687582 0.09687582
#>  [8,] 0.17750006 0.17750006 0.17750006
#>  [9,] 0.60368413 0.57925815 0.97054498
#> [10,] 0.18598632 0.97033116 0.97033116
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.6328491 0.01704143 0.01704143
#>  [2,] 1.0686219 0.96769261 0.53386264
#>  [3,] 0.3729635 1.56834107 1.21231765
#>  [4,] 2.2767016 1.16364342 1.81865352
#>  [5,] 0.1534754 0.31656625 0.31656625
#>  [6,] 1.6101621 0.03852917 2.69983121
#>  [7,] 0.2561493 0.25614926 0.25614926
#>  [8,] 0.2329622 0.23296220 0.23296220
#>  [9,] 1.0235122 1.02351223 0.69805601
#> [10,] 0.5859019 0.65846586 0.58590190
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.305473828 0.305473828 0.305473828
#>  [2,] 0.932243945 0.913498521 0.909356596
#>  [3,] 0.167620940 0.167620940 0.167620940
#>  [4,] 0.136362707 0.472767848 0.136362707
#>  [5,] 0.236457793 0.183129230 0.678797476
#>  [6,] 1.054354740 0.029667483 0.149710835
#>  [7,] 0.556171987 0.101790295 0.069821429
#>  [8,] 0.001083053 0.001083053 0.001083053
#>  [9,] 0.093009385 1.091734289 1.264046209
#> [10,] 0.002887774 0.002887774 0.232032799
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.43450242 1.24508941 1.04935347
#>  [2,] 0.14954542 0.14954542 0.14954542
#>  [3,] 0.08871447 0.08871447 0.08871447
#>  [4,] 0.92402561 0.77292334 0.92402561
#>  [5,] 0.26263231 0.26263231 0.26263231
#>  [6,] 0.59877110 0.59877110 0.59877110
#>  [7,] 0.18274097 0.18274097 0.18274097
#>  [8,] 0.01751393 0.55064249 0.01751393
#>  [9,] 0.22240655 0.22240655 0.67131780
#> [10,] 0.75982313 0.75982313 0.75982313
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>            [,1]        [,2]      [,3]
#>  [1,] 0.3398374 0.339837450 0.3398374
#>  [2,] 0.1404722 0.140472178 0.1404722
#>  [3,] 0.9538416 0.282588603 0.9538416
#>  [4,] 0.4708826 0.556093134 0.2233152
#>  [5,] 0.8579791 0.857979092 0.8579791
#>  [6,] 0.3384950 0.338494956 0.3225697
#>  [7,] 0.3884181 0.922132252 0.9221323
#>  [8,] 0.2809514 0.004760305 0.3729104
#>  [9,] 0.5139771 0.513977082 0.5139771
#> [10,] 0.2978054 0.214667874 0.3779164

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.288088307 0.174150818 0.604591856
#>  [2,] 0.594298727 0.747054531 0.022854975
#>  [3,] 0.283348960 0.080605639 0.080605639
#>  [4,] 0.004870368 0.004870368 0.004870368
#>  [5,] 1.464433679 0.669030059 0.673417237
#>  [6,] 0.579682116 0.096273483 0.200073725
#>  [7,] 0.610249988 0.344382233 0.610249988
#>  [8,] 0.278721369 0.278721369 0.278721369
#>  [9,] 0.057911376 1.062623429 1.422035497
#> [10,] 0.133249794 0.450478797 0.133249794
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.165041604 0.16504160 0.165041604
#>  [2,] 0.269530628 0.76535245 0.456172857
#>  [3,] 0.399040952 0.30711006 0.212574677
#>  [4,] 0.092136874 0.44213145 0.442131451
#>  [5,] 0.021882047 0.81514501 0.924126538
#>  [6,] 0.340267626 0.34026763 0.007515733
#>  [7,] 0.002253101 0.06084961 0.456320532
#>  [8,] 0.260211434 0.26021143 0.260211434
#>  [9,] 0.096903304 0.09690330 0.096903304
#> [10,] 0.160204137 1.31681162 1.316811621
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.41305717 0.40152810 0.41305717
#>  [2,] 0.11398210 0.11398210 0.11398210
#>  [3,] 0.17597457 0.43795953 0.65000475
#>  [4,] 0.37787257 0.94029983 0.10133178
#>  [5,] 1.03340630 0.09011970 0.46912458
#>  [6,] 0.10089156 0.07378819 0.39172361
#>  [7,] 0.01188837 0.01188837 0.01188837
#>  [8,] 0.39692986 0.03871148 0.39692986
#>  [9,] 0.14686480 0.56022595 0.57249432
#> [10,] 0.18807323 0.02456062 0.43828296
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55592358 0.95126266 0.13697833
#>  [2,] 0.12048952 0.60437525 0.18266905
#>  [3,] 0.74456891 0.53278176 0.29404637
#>  [4,] 0.53634989 0.87362177 0.55390030
#>  [5,] 0.54926671 0.08968877 0.13701657
#>  [6,] 0.29123430 0.29568336 0.22088845
#>  [7,] 0.40594618 0.23622295 0.19060360
#>  [8,] 0.08895539 0.08895539 0.08895539
#>  [9,] 0.17507154 0.38878387 0.02909609
#> [10,] 0.41233910 0.28295984 0.41233910

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.60764956 1.26159734 0.60764956
#>  [2,] 0.07428941 0.07428941 0.18107639
#>  [3,] 0.19285409 0.19285409 0.19285409
#>  [4,] 0.68479512 0.08496511 0.68479512
#>  [5,] 0.52973751 0.52180588 0.52180588
#>  [6,] 0.22241148 0.22241148 1.44711031
#>  [7,] 0.11585034 0.05038468 0.13027635
#>  [8,] 0.31505738 0.19488479 0.31505738
#>  [9,] 0.24936679 0.40701087 0.40701087
#> [10,] 0.15728128 0.15728128 0.03369683
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.221517394 0.06016175 0.346946737
#>  [2,] 0.208886991 0.48397230 0.052562032
#>  [3,] 1.544496313 0.61223038 1.544496313
#>  [4,] 0.232717982 0.26858812 0.457357008
#>  [5,] 0.004456786 0.68613999 0.004456786
#>  [6,] 0.217437954 0.44474954 0.444749538
#>  [7,] 1.158298805 2.48241103 0.238791876
#>  [8,] 0.323688235 0.32368823 0.275713674
#>  [9,] 0.025659450 0.17899701 0.025659450
#> [10,] 0.743576887 0.74357689 0.044566810
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]        [,2]      [,3]
#>  [1,] 0.34052641 0.440847836 0.4572884
#>  [2,] 0.05297758 0.310115909 0.3101159
#>  [3,] 0.39609514 0.396095143 0.3960951
#>  [4,] 0.09547656 0.175296752 0.1752968
#>  [5,] 0.44300880 0.675487211 0.1569040
#>  [6,] 0.45094728 1.289692369 1.9375585
#>  [7,] 0.26324401 0.331838609 0.1455853
#>  [8,] 0.02956718 0.148290445 0.1482904
#>  [9,] 0.51985570 0.007711438 0.5198557
#> [10,] 0.04113756 0.041137557 0.1066886
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.04009660 0.41308837 0.41308837
#>  [2,] 0.14062176 0.14062176 0.01563755
#>  [3,] 0.28487310 0.15952407 0.28487310
#>  [4,] 0.22780041 0.15330174 0.22780041
#>  [5,] 0.07952825 0.07952825 0.29139017
#>  [6,] 0.29988386 0.30792962 1.18491766
#>  [7,] 0.25870136 0.25870136 0.16778215
#>  [8,] 0.56693589 0.40139647 0.40139647
#>  [9,] 0.47601032 0.67856439 0.43273703
#> [10,] 1.43796393 1.26052779 0.74456618

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>              [,1]       [,2]       [,3]
#>  [1,] 0.041468198 0.04146820 0.04146820
#>  [2,] 0.723575719 0.72357572 0.72357572
#>  [3,] 0.772554209 1.07029409 0.77255421
#>  [4,] 0.502252432 0.14846088 1.65195780
#>  [5,] 0.056286133 0.05628613 0.02436355
#>  [6,] 0.151248272 0.50531450 0.15124827
#>  [7,] 0.003204724 0.28451287 0.28451287
#>  [8,] 0.015480927 0.43193446 0.61546529
#>  [9,] 0.539600440 0.44750897 0.41900848
#> [10,] 0.638917677 1.14678371 0.30689105
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.14530401 0.13257899 0.39632698
#>  [2,] 0.01981931 0.06636445 0.01981931
#>  [3,] 0.06832073 0.10480144 0.08179711
#>  [4,] 0.63646470 0.03809847 0.63646470
#>  [5,] 0.43526717 0.59325478 0.66881239
#>  [6,] 0.02528063 0.20930477 0.28391722
#>  [7,] 0.59999569 1.04591896 0.02052234
#>  [8,] 0.32602251 0.19898367 0.53530488
#>  [9,] 0.42012092 0.15296431 0.48718750
#> [10,] 0.05929961 0.05929961 0.05929961
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.09007746 0.09007746 0.09007746
#>  [2,] 0.83505071 0.38046227 0.48676787
#>  [3,] 0.14931538 0.14931538 0.14931538
#>  [4,] 0.51882896 0.55236131 0.55236131
#>  [5,] 1.23838944 0.64681839 1.23838944
#>  [6,] 0.28827925 0.28827925 0.13802713
#>  [7,] 0.40457152 0.40457152 0.40457152
#>  [8,] 0.32763541 0.06585085 0.32763541
#>  [9,] 0.23151956 0.20170571 0.33971232
#> [10,] 0.38998281 0.21020041 0.02289723
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>              [,1]       [,2]        [,3]
#>  [1,] 0.369294065 0.40469890 0.008382971
#>  [2,] 0.194608551 0.19460855 0.194608551
#>  [3,] 0.097693286 0.09769329 0.388150791
#>  [4,] 0.662801133 1.04085992 0.437567056
#>  [5,] 0.231478739 0.23147874 0.231478739
#>  [6,] 0.051449231 0.88915087 0.283890139
#>  [7,] 0.012505158 0.47039681 0.373742295
#>  [8,] 0.829876324 0.03447686 0.615427932
#>  [9,] 0.275617766 0.04832584 1.107650928
#> [10,] 0.009255035 0.07136233 0.009255035