Sample from parametrized ext. MO distributions

Draws \(n\) iid samples from a \(d\)-variate paramatrized ext. MO distribution.

Usage

rpextmo(
  n,
  d,
  a = 0,
  b = 0,
  gamma = 1,
  eta = NULL,
  family = c("Armageddon", "Poisson", "Pareto", "Exponential", "AlphaStable",
    "InverseGaussian", "Gamma"),
  method = c("MDCM", "LFM", "AM", "ESM")
)

Arguments

n

an integer for the number of samples.

d

an integer for the dimension.

a

a non-negative double for the killing-rate \(a\) of the Bernstein function.

b

a non-negative double for the drift \(b\) of the Bernstein function.

gamma

a position double for the scaling of the integral part of the Bernstein function.

eta

a numeric vector for the family's parameter.

family

a string indicating the parametrized family. Use "Armageddon" for the Armageddon family, "Poisson" for the Poisson family, "Pareto" for the Pareto family, "Exponential" for the Exponential family, "AlphaStable" for the \(\alpha\)-stable family, "InverseGaussian" for the Inverse-Gaussian family, "Gamma" for the Gamma family.

method

a string indicating which sampling algorithm should be used. Use "MDCM" for the Markovian death-set model, "LFM" for the Lévy-frailty model, "AM" for the Arnold model, and "ESM" for the exogenous shock model (in case of the Armageddon family, the algorithm is optimized to consider only finite shocks). We recommend using the ESM only for small dimensions; the AM can be used up until dimension \(30\).

Value

rpextmo returns a numeric matrix with \(n\) rows and \(d\) columns with the rows corresponding to iid distributed samples of a \(d\)-variate parametrized extendible Marshall–Olkin distribution with corresponding parameters.

Parametrization

A parametrized ext. MO distribution is a family of ext. MO distributions corresponding to Bernstein functions of the form $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ {[1 - e^{-ux}]} \nu{(\mathrm{d}u)} }, \quad x \geq 0 , $$ or $$ \psi{(x)} = 1_{\{ x > 0 \}} a + b x + \gamma \cdot \int_{0}^{\infty}{ \frac{x}{x + u} \sigma{(\mathrm{d}u)} }, \quad x \geq 0 , $$ where \(a, a \geq 0\) and \(\nu\), resp. \(\sigma\), are the Lévy measure, resp. Stieltjes measure, and \(\alpha > 0\), \(\beta > 0\), or \(nu \not\equiv 0\), resp. \(\sigma \not\equiv 0\).

The Lévy measure, resp. Stieltjes measure, take the forms listed in the following.

Armageddon family

We have \(nu = \sigma \equiv 0\) and Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x , \quad x \geq 0 , $$ see ConstantBernsteinFunction and LinearBernsteinFunction.

Poisson family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {[1 - e^{-\eta x}]}, \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see PoissonBernsteinFunction.

Pareto family

We have \(\eta \in \mathbb{R}^2\) with \(\eta_1, \eta_2 > 0\) and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta_{1} \eta_{2}^{\eta_{1}} \cdot u^{-\eta_{1}-1} 1_{\{ u > \eta_{2}\}} \mathrm{d}u , $$ see ParetoBernsteinFunction.

Exponential family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \frac{x}{x + \eta} , \quad x \geq 0 , $$ and Lévy measure $$ \nu{(\mathrm{d}u)} = \eta e^{-\eta u} \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \delta_{\{ \eta \}}{(\mathrm{d}u)} , $$ see ExponentialBernsteinFunction.

\(\alpha\)-stable family

We have \(\eta \in {(0, 1)}\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot x^{\eta} , \quad x \geq 0 , $$ Lévy measure, $$ \nu{(\mathrm{d}u)} = \frac{\eta}{\Gamma{(1 - \eta)}} \cdot u^{-\eta-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\eta \pi)}}{\pi} \cdot u^{\eta - 1} \mathrm{d}u , $$ see AlphaStableBernsteinFunction.

Inverse-Gaussian family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot {\left[ \sqrt{2 x + \eta^2} - \eta \right]}, \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = \frac{1}{ \sqrt{2 \pi} } \cdot \frac{ e^{-\frac{1}{2} \eta^2 u} }{ \sqrt{u^3} } \mathrm{d}u , $$ and Stieltjes measure $$ \sigma{(\mathrm{d}u)} = \frac{\sin{(\pi / 2)}}{\pi} \cdot \frac{\sqrt{2 u - \eta^2}}{u} 1_{\{ u > \eta^2 / 2 \}} \mathrm{d}u , $$ see InverseGaussianBernsteinFunction.

Gamma family

We have \(\eta > 0\), Bernstein function $$ \psi{(x)} = 1_{\{ x > 0\}} a + b x + \gamma \cdot \log{\left( 1 + \frac{x}{\eta} \right)} , \quad x \geq 0 , $$ Lévy measure $$ \nu{(\mathrm{d}u)} = e^{-\eta u} u^{-1} \mathrm{d}u , $$ and Stieljtes measure $$ \sigma{(\mathrm{d}u)} = u^{-1} 1_{\{ u > \eta \}} \mathrm{d}u , $$ see GammaBernsteinFunction.

Simulation algorithms

(General) exchangeable Marshall–Olkin sampling algorithms

The MDCM, AM, and ESM simulation algorithms for the exchangeable Marshall–Olkin distribution can be used; for this, the corresponding Bernstein function is passed to rextmo(). An exception is the ESM for the Armageddon family which uses an optimized version considering only finite shock times.

Lévy-frailty model

The Lévy-frailty model simulates the elements of the random vector as first-hitting times of a compound Poisson subordinator \(\Lambda\) into sets \((E_i, \infty)\) for iid unit exponential random variables. Here, the subordinator is a linear combination of a pure-drift subordinator, a pure-killing subordinator, and a pure-jump compound Poisson subordinator, i.e. $$ \Lambda_{t} = \infty \cdot 1_{\{ \epsilon > a t \}} + b t + \sum_{j=1}^{N_{\gamma t}} X_{j} , \quad t \geq 0, $$ where \(\epsilon\) is a unit exponential rv, \(N\) is a Poisson process, and \(X_{1}, X_{2}, \ldots\) are iid jumps from the corresponding jump distribution, see (see pp. 140 psqq. Mai and Scherer 2017) .

References

Mai J, Scherer M (2017). Simulating copulas: stochastic models, sampling algorithms and applications, Series in Quantitative Finance, 2 edition. World Scientific. doi:10.1142/10265 .

See also

Examples

## Armageddon
rpextmo(10, 3, a = 0.2, b = 0.5)
#>             [,1]        [,2]      [,3]
#>  [1,] 4.84696406 0.003079934 0.6397086
#>  [2,] 0.87013364 2.826124685 2.3834746
#>  [3,] 0.05059676 1.189277019 0.9696219
#>  [4,] 1.93916104 0.034903594 1.3845794
#>  [5,] 3.50428950 1.664561052 1.3020882
#>  [6,] 0.15305381 0.501533689 2.3624253
#>  [7,] 1.01941854 0.151933373 0.6932293
#>  [8,] 0.80198824 0.256082174 1.1052218
#>  [9,] 0.34764650 1.843778809 0.9185155
#> [10,] 0.43034808 0.430348080 0.4303481
rpextmo(10, 3, a = 1)      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4537644 0.4537644 0.4537644
#>  [2,] 1.9978247 1.9978247 1.9978247
#>  [3,] 0.8705878 0.8705878 0.8705878
#>  [4,] 0.2670871 0.2670871 0.2670871
#>  [5,] 2.1741897 2.1741897 2.1741897
#>  [6,] 1.8515082 1.8515082 1.8515082
#>  [7,] 3.2527188 3.2527188 3.2527188
#>  [8,] 0.9820625 0.9820625 0.9820625
#>  [9,] 0.6995112 0.6995112 0.6995112
#> [10,] 0.3902669 0.3902669 0.3902669
rpextmo(10, 3, b = 1)      ## independence
#>             [,1]        [,2]      [,3]
#>  [1,] 0.79933225 0.294649656 0.1250811
#>  [2,] 0.66393955 1.128389197 1.9581605
#>  [3,] 0.21449801 0.038205571 0.6002769
#>  [4,] 0.33818782 0.604351732 0.7479987
#>  [5,] 0.00804703 0.003437873 0.1096467
#>  [6,] 0.49825923 0.784171041 1.9517075
#>  [7,] 1.06769320 2.391962604 0.5096338
#>  [8,] 0.22219829 1.674550665 1.6769706
#>  [9,] 1.22535908 0.979601626 1.1179401
#> [10,] 0.20243473 1.572801685 0.7172216

rpextmo(10, 3, a = 0.2, b = 0.5, method = "ESM")
#>            [,1]       [,2]       [,3]
#>  [1,] 3.3422577 0.69001361 4.69173578
#>  [2,] 2.0999858 0.94755992 0.06617116
#>  [3,] 1.8978480 2.82539579 1.94514754
#>  [4,] 0.3962758 0.05826789 0.71154932
#>  [5,] 2.9888478 0.17977489 0.27207411
#>  [6,] 1.9146054 1.91460536 1.48359793
#>  [7,] 0.1560561 0.72459446 4.02833052
#>  [8,] 0.5728799 3.03981272 0.68633088
#>  [9,] 0.1243341 0.12433410 0.12433410
#> [10,] 1.1411391 4.18772972 0.20642865
rpextmo(10, 3, a = 1, method = "ESM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.36802683 0.36802683 0.36802683
#>  [2,] 1.43175302 1.43175302 1.43175302
#>  [3,] 0.02444364 0.02444364 0.02444364
#>  [4,] 0.01389856 0.01389856 0.01389856
#>  [5,] 0.06786033 0.06786033 0.06786033
#>  [6,] 0.30589156 0.30589156 0.30589156
#>  [7,] 0.36721432 0.36721432 0.36721432
#>  [8,] 1.69070404 1.69070404 1.69070404
#>  [9,] 0.11424527 0.11424527 0.11424527
#> [10,] 0.78800193 0.78800193 0.78800193
rpextmo(10, 3, b = 1, method = "ESM")      ## independence
#>             [,1]        [,2]       [,3]
#>  [1,] 0.04687067 0.927005312 0.95930204
#>  [2,] 1.86271417 1.797928678 1.13546574
#>  [3,] 2.06414716 0.009943752 0.20405217
#>  [4,] 1.64586819 0.852672397 0.17044957
#>  [5,] 0.91423636 0.198963706 0.03227045
#>  [6,] 0.38829912 0.433557968 1.37731941
#>  [7,] 0.24146707 0.027134103 0.10568179
#>  [8,] 2.64946193 2.526120522 1.78398215
#>  [9,] 0.30694454 5.051515325 0.18495807
#> [10,] 1.83778850 0.605925967 0.17724860

rpextmo(10, 3, a = 0.2, b = 0.5, method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3645882 1.3645882 0.8812951
#>  [2,] 0.8232749 3.0760127 2.2773590
#>  [3,] 2.1833711 0.4694165 0.5323077
#>  [4,] 2.7480343 0.6633109 0.2062201
#>  [5,] 1.5344727 1.2640669 0.1448433
#>  [6,] 1.6703235 0.8432706 1.6703235
#>  [7,] 2.0157225 2.3354318 0.2233405
#>  [8,] 2.9405579 4.7152120 0.6173184
#>  [9,] 0.7017299 0.4769886 0.3157100
#> [10,] 0.5167216 1.4492920 3.5183571
rpextmo(10, 3, a = 1, method = "LFM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 2.0197855 2.0197855 2.0197855
#>  [2,] 0.2241251 0.2241251 0.2241251
#>  [3,] 0.6179966 0.6179966 0.6179966
#>  [4,] 0.9256871 0.9256871 0.9256871
#>  [5,] 0.6003719 0.6003719 0.6003719
#>  [6,] 0.2227200 0.2227200 0.2227200
#>  [7,] 0.3133145 0.3133145 0.3133145
#>  [8,] 0.9591248 0.9591248 0.9591248
#>  [9,] 0.1248450 0.1248450 0.1248450
#> [10,] 2.1771217 2.1771217 2.1771217
rpextmo(10, 3, b = 1, method = "LFM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.19549328 1.3248045 0.03181234
#>  [2,] 0.36636656 0.1585201 0.04459026
#>  [3,] 0.61340601 1.8801888 0.99950094
#>  [4,] 2.58164974 2.5713632 0.78441546
#>  [5,] 4.30848195 0.1684687 2.59953545
#>  [6,] 3.75712602 0.7801753 0.88784986
#>  [7,] 0.05441596 0.6300294 2.20086199
#>  [8,] 0.09261105 0.3137999 2.01941999
#>  [9,] 0.81082392 0.3210410 0.10473871
#> [10,] 0.63456764 0.4264079 1.14230684

rpextmo(10, 3, a = 0.2, b = 0.5, method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 4.57682353 0.37315589 0.46301766
#>  [2,] 0.08521225 2.15202942 2.15202942
#>  [3,] 0.27826104 3.47841347 1.53531695
#>  [4,] 0.39463206 0.39463206 0.39463206
#>  [5,] 0.21916398 0.21916398 0.21916398
#>  [6,] 3.09950898 1.35774646 1.10060892
#>  [7,] 2.43922151 0.58351909 0.24862177
#>  [8,] 0.39937024 2.00369990 0.37282838
#>  [9,] 3.44910119 0.69227683 2.31260687
#> [10,] 0.05177242 0.05177242 0.05177242
rpextmo(10, 3, a = 1, method = "MDCM")      ## comonotone
#>              [,1]        [,2]        [,3]
#>  [1,] 0.861431578 0.861431578 0.861431578
#>  [2,] 2.490638891 2.490638891 2.490638891
#>  [3,] 1.608765976 1.608765976 1.608765976
#>  [4,] 0.002849338 0.002849338 0.002849338
#>  [5,] 0.245437361 0.245437361 0.245437361
#>  [6,] 1.791703643 1.791703643 1.791703643
#>  [7,] 4.070400884 4.070400884 4.070400884
#>  [8,] 1.126726505 1.126726505 1.126726505
#>  [9,] 0.644740581 0.644740581 0.644740581
#> [10,] 1.074577736 1.074577736 1.074577736
rpextmo(10, 3, b = 1, method = "MDCM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 1.4628214 0.99243593 1.6301628
#>  [2,] 0.5330587 0.19557276 2.3492531
#>  [3,] 5.3631221 1.57752125 1.0807759
#>  [4,] 1.1328879 0.08363711 0.5438171
#>  [5,] 1.7502170 0.12836185 1.9890085
#>  [6,] 2.0316784 0.35194902 0.4334137
#>  [7,] 0.8785628 0.20009845 0.6353904
#>  [8,] 0.4646680 0.08313862 0.1174632
#>  [9,] 0.1348595 0.34225704 1.0162989
#> [10,] 2.3876815 0.26000150 1.0277979

rpextmo(10, 3, a = 0.2, b = 0.5, method = "AM")
#>              [,1]      [,2]      [,3]
#>  [1,] 0.268132804 0.8240713 0.8240713
#>  [2,] 0.476545872 1.2068458 1.2068458
#>  [3,] 0.262612085 0.1749027 1.6868291
#>  [4,] 0.623258191 0.7796512 0.6784471
#>  [5,] 0.111796767 3.6016959 3.6016959
#>  [6,] 0.009049397 2.1035839 2.5901175
#>  [7,] 0.198308472 2.4207662 2.4207662
#>  [8,] 0.470896588 0.4708966 0.3743182
#>  [9,] 3.576561712 1.8251936 1.8289726
#> [10,] 1.522443594 1.5224436 1.5224436
rpextmo(10, 3, a = 1, method = "AM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.4676639 0.4676639 0.4676639
#>  [2,] 0.4771550 0.4771550 0.4771550
#>  [3,] 1.7994273 1.7994273 1.7994273
#>  [4,] 2.4706893 2.4706893 2.4706893
#>  [5,] 0.9305721 0.9305721 0.9305721
#>  [6,] 0.1204414 0.1204414 0.1204414
#>  [7,] 1.7138655 1.7138655 1.7138655
#>  [8,] 0.8180815 0.8180815 0.8180815
#>  [9,] 0.9131796 0.9131796 0.9131796
#> [10,] 0.1321979 0.1321979 0.1321979
rpextmo(10, 3, b = 1, method = "AM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 0.2738141 1.24577827 1.10850083
#>  [2,] 1.3023405 0.80226376 0.64489150
#>  [3,] 0.4380295 2.61362155 0.01780944
#>  [4,] 2.7455043 0.04271667 1.44386800
#>  [5,] 1.1757734 1.53750217 0.17419998
#>  [6,] 0.5708655 0.22845009 0.36865043
#>  [7,] 0.1892345 0.84000507 0.15760431
#>  [8,] 1.9303909 0.45445120 0.35565883
#>  [9,] 0.1964722 1.97103586 0.42687499
#> [10,] 0.7939952 0.07218586 1.13485873

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.7383208 0.68036758 0.1211298
#>  [2,] 0.3729020 0.04467767 2.0689100
#>  [3,] 5.8154923 2.31009310 3.7809084
#>  [4,] 3.2343615 2.47333090 0.9178670
#>  [5,] 0.1445212 2.29154102 2.0262437
#>  [6,] 1.4018089 1.81774962 0.9309521
#>  [7,] 1.8819409 0.60406153 0.9223950
#>  [8,] 0.7355216 1.16925961 0.9052331
#>  [9,] 0.6543818 7.21695073 4.3885804
#> [10,] 0.1310771 0.13107712 0.1310771
rpextmo(10, 3, a = 1, family = "Armageddon")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 1.0254636 1.0254636 1.0254636
#>  [2,] 2.2497428 2.2497428 2.2497428
#>  [3,] 0.5100132 0.5100132 0.5100132
#>  [4,] 0.8325356 0.8325356 0.8325356
#>  [5,] 3.9602182 3.9602182 3.9602182
#>  [6,] 0.3087031 0.3087031 0.3087031
#>  [7,] 0.5058701 0.5058701 0.5058701
#>  [8,] 4.0223784 4.0223784 4.0223784
#>  [9,] 1.8347566 1.8347566 1.8347566
#> [10,] 1.6445850 1.6445850 1.6445850
rpextmo(10, 3, b = 1, family = "Armageddon")      ## independence
#>             [,1]       [,2]      [,3]
#>  [1,] 2.21804334 0.67252135 1.5759440
#>  [2,] 0.44825665 2.38124677 1.7495824
#>  [3,] 1.40000560 0.29865254 1.0166123
#>  [4,] 0.56334499 1.59113404 0.2596070
#>  [5,] 0.73092424 0.69319010 0.2237879
#>  [6,] 0.03266545 0.19833915 2.3820922
#>  [7,] 0.18938044 0.10168752 0.9226974
#>  [8,] 2.54626158 0.02162302 2.6405162
#>  [9,] 1.16661967 1.98481159 2.8051517
#> [10,] 0.22783652 1.60642230 0.2977621

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "ESM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.9907222 0.8894966 3.1009940
#>  [2,] 1.2552852 2.5855944 0.2649401
#>  [3,] 0.9113357 0.9113357 0.9113357
#>  [4,] 2.8251279 3.4662454 5.3573492
#>  [5,] 2.3782731 1.3848973 2.5068134
#>  [6,] 1.0568048 2.6601687 2.6601687
#>  [7,] 4.6306987 3.6664051 2.5924963
#>  [8,] 2.9379712 2.6664342 0.1301347
#>  [9,] 1.2048114 4.3757817 1.4612366
#> [10,] 5.3793615 0.4247765 1.8988235
rpextmo(10, 3, a = 1, family = "Armageddon", method = "ESM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.9182297 0.9182297 0.9182297
#>  [2,] 0.8647747 0.8647747 0.8647747
#>  [3,] 0.6287158 0.6287158 0.6287158
#>  [4,] 2.8616955 2.8616955 2.8616955
#>  [5,] 0.9518616 0.9518616 0.9518616
#>  [6,] 0.1491793 0.1491793 0.1491793
#>  [7,] 0.2466800 0.2466800 0.2466800
#>  [8,] 0.3373677 0.3373677 0.3373677
#>  [9,] 2.0794318 2.0794318 2.0794318
#> [10,] 1.8140350 1.8140350 1.8140350
rpextmo(10, 3, b = 1, family = "Armageddon", method = "ESM")      ## independence
#>              [,1]      [,2]       [,3]
#>  [1,] 0.847845608 0.0135050 0.36797203
#>  [2,] 2.645776557 1.4991199 0.12913484
#>  [3,] 0.696852234 1.4670748 1.51250086
#>  [4,] 0.443412614 2.4131168 0.25325142
#>  [5,] 0.818908993 0.6076187 1.35675166
#>  [6,] 0.156636259 0.2515792 1.00029197
#>  [7,] 0.121688502 0.4789348 4.07187185
#>  [8,] 0.005144272 0.3145388 0.57178831
#>  [9,] 1.952933401 1.1536425 0.07572778
#> [10,] 0.054485100 1.3722523 0.17576443

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 2.7897430 0.7287540 0.9303866
#>  [2,] 0.8849436 2.1800735 2.1800735
#>  [3,] 0.3089307 0.3089307 0.3089307
#>  [4,] 0.4473862 0.4473862 0.4473862
#>  [5,] 1.0545696 0.0233782 1.3572546
#>  [6,] 0.3303897 2.4299667 2.2185251
#>  [7,] 0.8825237 0.1896363 0.1083827
#>  [8,] 0.6738414 1.9761097 0.3057555
#>  [9,] 0.6685258 0.6685258 0.6685258
#> [10,] 0.4052073 0.8936162 0.5418658
rpextmo(10, 3, a = 1, family = "Armageddon", method = "LFM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.59918748 0.59918748 0.59918748
#>  [2,] 0.78557047 0.78557047 0.78557047
#>  [3,] 3.56486995 3.56486995 3.56486995
#>  [4,] 1.01624357 1.01624357 1.01624357
#>  [5,] 0.02918399 0.02918399 0.02918399
#>  [6,] 0.66488173 0.66488173 0.66488173
#>  [7,] 0.21823752 0.21823752 0.21823752
#>  [8,] 1.89960320 1.89960320 1.89960320
#>  [9,] 1.00235828 1.00235828 1.00235828
#> [10,] 1.22094751 1.22094751 1.22094751
rpextmo(10, 3, b = 1, family = "Armageddon", method = "LFM")      ## independence
#>             [,1]      [,2]       [,3]
#>  [1,] 0.07872534 0.1880067 0.59385809
#>  [2,] 0.20482411 0.1045839 1.15581234
#>  [3,] 0.28103700 2.8816220 4.70312639
#>  [4,] 1.87139997 1.5077261 1.46637333
#>  [5,] 1.33295799 0.8927885 0.63793239
#>  [6,] 0.48677318 1.5924631 0.38777656
#>  [7,] 1.41402597 0.2130039 0.06568895
#>  [8,] 0.65189670 1.8129012 1.19896524
#>  [9,] 0.58441731 0.6220814 0.87599947
#> [10,] 0.31045214 2.9035727 0.21258835

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.73785334 0.73065598 1.32352092
#>  [2,] 0.13989122 0.09786354 0.02554489
#>  [3,] 2.00924478 2.69720733 0.08804492
#>  [4,] 2.44860443 1.89989912 1.35193697
#>  [5,] 0.01865419 0.01865419 0.01865419
#>  [6,] 1.00612700 1.00612700 1.00612700
#>  [7,] 0.48697726 0.48697726 0.47826388
#>  [8,] 0.03655302 0.21169721 0.21169721
#>  [9,] 1.16113964 0.43650051 0.23555944
#> [10,] 0.46962234 1.65442658 0.86001988
rpextmo(10, 3, a = 1, family = "Armageddon", method = "MDCM")      ## comonotone
#>            [,1]      [,2]      [,3]
#>  [1,] 0.1685715 0.1685715 0.1685715
#>  [2,] 0.3298743 0.3298743 0.3298743
#>  [3,] 1.0168984 1.0168984 1.0168984
#>  [4,] 0.5825360 0.5825360 0.5825360
#>  [5,] 1.2958196 1.2958196 1.2958196
#>  [6,] 0.6270296 0.6270296 0.6270296
#>  [7,] 5.0256793 5.0256793 5.0256793
#>  [8,] 2.9610276 2.9610276 2.9610276
#>  [9,] 0.2316858 0.2316858 0.2316858
#> [10,] 1.9691536 1.9691536 1.9691536
rpextmo(10, 3, b = 1, family = "Armageddon", method = "MDCM")      ## independence
#>            [,1]       [,2]      [,3]
#>  [1,] 0.8828871 0.13998043 5.0634733
#>  [2,] 0.5846351 0.46648464 0.4893888
#>  [3,] 1.0961900 0.20252730 0.7830626
#>  [4,] 0.2066092 0.36333712 0.5828560
#>  [5,] 0.3026111 3.57558619 0.4565346
#>  [6,] 1.6576565 0.09170593 1.3084592
#>  [7,] 1.3132463 0.26502864 0.7504237
#>  [8,] 1.4831892 0.28192544 1.9270301
#>  [9,] 4.2254664 1.72053249 1.6468397
#> [10,] 0.3380138 3.46141338 0.3664009

rpextmo(10, 3, a = 0.2, b = 0.5, family = "Armageddon", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 1.1857489 3.01728703 0.2568616
#>  [2,] 1.7426265 0.92696957 1.7426265
#>  [3,] 1.0307197 1.03071967 0.4228464
#>  [4,] 0.1559953 2.64386861 2.1087822
#>  [5,] 4.7358135 0.33919428 4.4008942
#>  [6,] 0.4732927 1.30267978 1.3071618
#>  [7,] 0.3128951 1.61405969 1.2079903
#>  [8,] 0.5099509 0.04575637 1.8923252
#>  [9,] 0.9190853 0.31358075 1.6391531
#> [10,] 0.1758627 0.38613037 2.6851733
rpextmo(10, 3, a = 1, family = "Armageddon", method = "AM")      ## comonotone
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02660791 0.02660791 0.02660791
#>  [2,] 0.21137847 0.21137847 0.21137847
#>  [3,] 0.35161489 0.35161489 0.35161489
#>  [4,] 1.10386404 1.10386404 1.10386404
#>  [5,] 0.18179776 0.18179776 0.18179776
#>  [6,] 0.02888418 0.02888418 0.02888418
#>  [7,] 0.13708795 0.13708795 0.13708795
#>  [8,] 0.97973197 0.97973197 0.97973197
#>  [9,] 0.98629665 0.98629665 0.98629665
#> [10,] 0.03513016 0.03513016 0.03513016
rpextmo(10, 3, b = 1, family = "Armageddon", method = "AM")      ## independence
#>            [,1]       [,2]       [,3]
#>  [1,] 1.2942757 0.16520561 1.79904835
#>  [2,] 2.2234953 2.71493983 1.74914511
#>  [3,] 1.1401017 0.74012764 0.25133051
#>  [4,] 2.1420880 0.01141846 0.69225701
#>  [5,] 3.3207557 0.08519066 0.03265196
#>  [6,] 0.5178563 2.24079154 0.64429330
#>  [7,] 2.1059478 2.36264347 0.21308518
#>  [8,] 0.1780055 1.07345144 1.25085098
#>  [9,] 1.8601947 4.79353776 0.31725126
#> [10,] 0.8229006 0.13614514 1.80047614

## Poisson
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson")
#>               [,1]       [,2]      [,3]
#>  [1,] 1.3755200790 1.24547103 1.3755201
#>  [2,] 0.0628061339 0.06280613 0.7074987
#>  [3,] 0.6614213177 2.13830734 1.1671661
#>  [4,] 0.0007848641 2.07028915 0.3407132
#>  [5,] 0.9017420719 0.57152055 0.5715206
#>  [6,] 0.2865061101 1.10154638 0.2865061
#>  [7,] 0.1794800621 0.17948006 0.1794801
#>  [8,] 0.0909415394 0.51154503 0.5115450
#>  [9,] 1.9040767043 2.63009073 0.3264137
#> [10,] 1.2842292773 2.47312796 0.6634194
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.55460554 0.31005971 1.25356220
#>  [2,] 0.10085910 0.03205667 0.03125744
#>  [3,] 0.16765957 0.17593654 0.56220020
#>  [4,] 0.01458676 0.01458676 0.17531297
#>  [5,] 3.39765117 1.26719400 1.66136044
#>  [6,] 0.39948546 0.08195714 0.67622597
#>  [7,] 6.37626509 1.45794952 2.25485138
#>  [8,] 0.45094532 0.12838655 1.26380518
#>  [9,] 0.06831046 0.06831046 0.06831046
#> [10,] 0.50934862 0.55053002 0.50934862
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "LFM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05673191 0.42057670 0.05673191
#>  [2,] 1.63063556 1.21390307 0.33457441
#>  [3,] 0.44264173 0.44264173 0.44264173
#>  [4,] 0.14042962 0.36171427 0.18704397
#>  [5,] 0.22278134 0.57586356 0.22278134
#>  [6,] 1.20772714 0.44193524 1.13165235
#>  [7,] 0.04647322 0.04647322 1.50059120
#>  [8,] 0.49931192 1.05936210 0.37791300
#>  [9,] 0.14832808 0.81021389 0.53848438
#> [10,] 1.18245164 0.74726756 0.13408731
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 1.20570012 0.25977029 0.25977029
#>  [2,] 0.02474247 3.43976418 0.02474247
#>  [3,] 1.11728525 0.94602849 1.26191393
#>  [4,] 0.10619124 0.05805109 0.05805109
#>  [5,] 1.88179174 1.00379248 2.48498449
#>  [6,] 0.28149583 0.07492805 0.28381154
#>  [7,] 0.19281053 0.19281053 0.19281053
#>  [8,] 0.58858276 0.98205719 0.58858276
#>  [9,] 0.75167291 0.75167291 0.06273600
#> [10,] 0.08231738 0.40082413 0.22888895
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Poisson", method = "AM")
#>            [,1]       [,2]       [,3]
#>  [1,] 0.1262111 1.62472690 2.50777450
#>  [2,] 0.1737051 3.12298800 0.74126842
#>  [3,] 0.2171630 0.06066331 0.06066331
#>  [4,] 1.7329795 0.04134331 2.05606200
#>  [5,] 0.4840946 0.59456052 0.76367300
#>  [6,] 0.2590768 0.16068226 0.43170015
#>  [7,] 0.5246598 1.03871904 1.03871904
#>  [8,] 2.2312210 1.26913209 0.39903811
#>  [9,] 0.1924755 0.19247545 0.13474504
#> [10,] 0.5774618 0.92133105 0.03433012

## Pareto
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.32840174 0.14530629 2.58021214
#>  [2,] 2.20796147 3.45218433 2.87444584
#>  [3,] 4.08994733 1.70209599 2.22189634
#>  [4,] 0.24118658 1.95147919 1.00082158
#>  [5,] 0.26429375 0.26195588 0.85071037
#>  [6,] 1.98244729 1.37573798 1.72073074
#>  [7,] 1.01645074 0.56896360 0.75726813
#>  [8,] 0.50394667 1.73260833 4.26862285
#>  [9,] 1.55761952 1.32683924 0.45752927
#> [10,] 0.09854436 0.09854436 0.09854436
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.53360216 0.93800575 0.62939913
#>  [2,] 1.03591841 0.98042711 1.03591841
#>  [3,] 0.17060385 2.44026215 1.11269981
#>  [4,] 0.01158474 1.91623828 0.49748001
#>  [5,] 0.12625840 2.01863758 1.21081912
#>  [6,] 0.04816392 0.04816392 0.60498869
#>  [7,] 1.24401767 1.24401767 0.79603244
#>  [8,] 1.10991634 0.48268806 0.01301188
#>  [9,] 1.95773288 1.17127753 1.95773288
#> [10,] 2.60879808 1.22230540 1.56046499
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "LFM")
#>            [,1]      [,2]      [,3]
#>  [1,] 1.3068423 1.4714333 1.4714333
#>  [2,] 1.7563892 1.7110349 1.7563892
#>  [3,] 1.8602587 4.1258428 3.7500381
#>  [4,] 0.3116787 0.8388847 0.8388847
#>  [5,] 1.9163602 3.5871240 1.1375826
#>  [6,] 1.2844797 2.1836782 0.5411001
#>  [7,] 0.3595786 0.3595786 0.3595786
#>  [8,] 0.2962108 1.3199997 2.8153963
#>  [9,] 0.6896261 0.6896261 0.1827636
#> [10,] 0.3775753 0.3775753 0.3775753
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "MDCM")
#>            [,1]       [,2]       [,3]
#>  [1,] 2.9636605 1.68739427 2.96366045
#>  [2,] 1.4533332 0.54293234 1.39104558
#>  [3,] 0.1228306 0.12283058 0.12283058
#>  [4,] 1.4615939 1.46159389 0.04861268
#>  [5,] 0.1118398 3.15036010 1.10125642
#>  [6,] 1.6716599 0.51420336 0.68366185
#>  [7,] 0.8984828 2.17024614 2.17024614
#>  [8,] 1.1914857 0.05213815 1.52521855
#>  [9,] 0.3822929 0.27027107 0.07952827
#> [10,] 7.2443670 3.14813584 0.87042253
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = c(0.5, 1e-4), family = "Pareto", method = "AM")
#>             [,1]       [,2]       [,3]
#>  [1,] 2.41017663 0.82445106 0.32265296
#>  [2,] 0.45509826 0.72415949 1.12488021
#>  [3,] 1.77548431 0.03293949 0.06213427
#>  [4,] 0.83135154 0.35730737 0.83135154
#>  [5,] 4.38155283 3.78379142 0.06938788
#>  [6,] 1.78313059 0.16212941 1.78313059
#>  [7,] 0.12626581 2.27010670 0.65127870
#>  [8,] 0.65047530 1.31633509 0.51530437
#>  [9,] 0.05247598 0.09967943 1.67324899
#> [10,] 0.44577429 0.44577429 0.44577429

## Exponential
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.23125073 0.23125073 0.23125073
#>  [2,] 0.87564444 0.87564444 0.11747156
#>  [3,] 0.07140392 0.07140392 0.07140392
#>  [4,] 1.40017158 1.40017158 0.86422472
#>  [5,] 0.53432327 0.53432327 0.53432327
#>  [6,] 0.91687293 0.11652919 0.91687293
#>  [7,] 0.24383570 0.24383570 0.24383570
#>  [8,] 0.64861876 0.20882398 0.52179711
#>  [9,] 0.03525460 0.62259589 0.03525460
#> [10,] 0.12945220 0.12945220 0.12945220
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "ESM")
#>             [,1]      [,2]      [,3]
#>  [1,] 0.35189700 1.5598555 0.8618617
#>  [2,] 0.52263569 0.5226357 0.5226357
#>  [3,] 0.34635612 0.3463561 0.4259338
#>  [4,] 0.25223146 0.2522315 0.2522315
#>  [5,] 1.48243177 0.7422964 2.0586555
#>  [6,] 1.43688198 1.5616676 0.6635521
#>  [7,] 0.23685286 0.2368529 0.2368529
#>  [8,] 0.43264788 0.2375829 0.2784999
#>  [9,] 0.50973607 1.1978066 0.1042345
#> [10,] 0.06231659 0.1580597 0.3047345
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "LFM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.01391771 0.921281189 0.01391771
#>  [2,] 0.83030685 1.079178449 0.89650947
#>  [3,] 0.42684015 0.003597897 1.84393582
#>  [4,] 0.08585426 0.184289604 0.54951160
#>  [5,] 0.04048471 1.155706500 1.15570650
#>  [6,] 0.46082105 0.460821054 0.46082105
#>  [7,] 0.52300929 1.455546797 0.02388862
#>  [8,] 0.16042831 0.160428305 0.03281425
#>  [9,] 1.23311245 0.306589108 1.23311245
#> [10,] 0.37191491 0.371914907 0.37191491
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.53669251 0.65689218 0.65689218
#>  [2,] 0.25617750 0.56918458 0.56918458
#>  [3,] 0.52190264 1.14490745 1.70036671
#>  [4,] 0.20797755 0.20797755 0.20797755
#>  [5,] 0.21640159 0.06724504 0.40628571
#>  [6,] 0.39733175 0.39733175 0.31723109
#>  [7,] 0.21685755 0.45922795 1.14938112
#>  [8,] 0.20452851 0.20452851 0.35860150
#>  [9,] 0.05007605 0.05007605 0.05007605
#> [10,] 0.11564105 0.24483410 0.11564105
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Exponential", method = "AM")
#>            [,1]       [,2]      [,3]
#>  [1,] 0.4337584 0.03026645 0.4337584
#>  [2,] 0.1417075 0.14170749 0.1417075
#>  [3,] 0.1072917 0.09786024 0.1072917
#>  [4,] 0.5629172 0.56291720 0.5629172
#>  [5,] 0.2030732 0.20307319 0.2030732
#>  [6,] 0.1992486 0.19924863 0.1841686
#>  [7,] 0.6100175 0.50101009 0.6100175
#>  [8,] 0.6689052 0.18318126 1.5849307
#>  [9,] 0.1029650 0.10296497 0.1029650
#> [10,] 0.1242213 0.30091812 0.3009181

## Alpha-Stable
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable")
#>               [,1]       [,2]      [,3]
#>  [1,] 0.2306833652 0.85837457 0.8583746
#>  [2,] 0.1793106037 0.59062199 0.1387754
#>  [3,] 0.1378441708 0.13784417 0.1378442
#>  [4,] 0.5456585158 0.24770337 0.2768301
#>  [5,] 0.0006501076 0.07099250 0.6072851
#>  [6,] 0.2061907501 0.35278426 0.3527843
#>  [7,] 0.2060129858 0.70575668 0.1203212
#>  [8,] 0.3698844122 0.05970931 0.3698844
#>  [9,] 0.1216617318 0.12166173 0.1099424
#> [10,] 0.2571171631 0.25274701 0.0703179
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.02885792 0.02321129 0.02885792
#>  [2,] 0.08800975 0.77419418 0.10255203
#>  [3,] 1.27551013 0.55833034 1.31776434
#>  [4,] 1.40240442 1.33363287 0.67683201
#>  [5,] 0.04753473 0.04753473 0.04753473
#>  [6,] 0.30074170 0.70999021 0.26438617
#>  [7,] 0.22057575 0.67367226 0.67367226
#>  [8,] 1.39913150 1.41013858 1.41013858
#>  [9,] 0.35052955 0.27947031 0.46054900
#> [10,] 0.03609187 0.03609187 0.03609187
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.06794544 0.004301405 0.06794544
#>  [2,] 0.48502562 1.043133149 0.78381399
#>  [3,] 0.32488056 0.578770497 0.21573359
#>  [4,] 0.09321111 0.093211109 0.09321111
#>  [5,] 0.60792027 1.254660178 1.14881694
#>  [6,] 0.56848185 0.259034775 0.12134294
#>  [7,] 0.06738712 0.312638118 0.06738712
#>  [8,] 0.46969808 0.448961863 0.76426628
#>  [9,] 0.94843628 0.716994920 0.59481219
#> [10,] 0.09784778 0.097847785 0.09784778
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "AlphaStable", method = "AM")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.065968118 0.065968118 0.015320054
#>  [2,] 0.124104371 0.315093800 0.083946602
#>  [3,] 0.008993638 0.008993638 0.008993638
#>  [4,] 0.200252940 0.200252940 0.108601618
#>  [5,] 0.014786137 0.014786137 0.014786137
#>  [6,] 0.703359611 0.913082487 0.116898777
#>  [7,] 0.176370612 0.506543168 0.936412153
#>  [8,] 0.015752142 0.257247551 0.015752142
#>  [9,] 0.024613107 0.024613107 0.024613107
#> [10,] 0.405883562 0.045198518 0.377082300

## Inverse Gaussian
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.05329191 0.05329191 0.05329191
#>  [2,] 1.03679103 0.30245345 0.30245345
#>  [3,] 0.04051969 0.55514626 0.04051969
#>  [4,] 0.41298338 0.12471917 1.30381459
#>  [5,] 0.18459080 0.38763640 0.58837126
#>  [6,] 0.17247709 0.17247709 0.17247709
#>  [7,] 0.07504893 0.07504893 0.07504893
#>  [8,] 0.27703027 0.02347585 0.02347585
#>  [9,] 0.65359984 0.02179437 0.32060289
#> [10,] 0.93404259 0.37602966 0.50556993
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "ESM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.17298554 0.27955771 0.28731298
#>  [2,] 0.58431270 1.39573445 0.25305676
#>  [3,] 0.33421219 0.33421219 0.05119279
#>  [4,] 0.19208145 0.19208145 0.19208145
#>  [5,] 0.03753679 0.41596186 0.07935961
#>  [6,] 0.34586369 0.37601071 0.21518223
#>  [7,] 0.22675247 0.22720028 0.59157256
#>  [8,] 0.28551735 0.54293589 0.25602404
#>  [9,] 0.51747145 0.31615760 0.31615760
#> [10,] 0.48142285 0.04309388 0.48142285
rpextmo(
 10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "MDCM")
#>             [,1]       [,2]       [,3]
#>  [1,] 0.56177339 0.08342832 0.36203951
#>  [2,] 0.53667540 1.02892398 0.13859834
#>  [3,] 0.61568270 0.50948936 0.14612122
#>  [4,] 0.05203537 0.34039790 0.89211641
#>  [5,] 0.06325944 0.11751132 0.39939869
#>  [6,] 0.07803089 0.33907938 0.33907938
#>  [7,] 0.09098858 0.09098858 1.12131401
#>  [8,] 0.01664286 1.49190346 0.06375070
#>  [9,] 1.64772193 1.84951061 0.06027568
#> [10,] 0.53903965 0.03877889 0.13516461
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "InverseGaussian", method = "AM")
#>             [,1]       [,2]      [,3]
#>  [1,] 0.26820072 1.15678499 0.7198571
#>  [2,] 0.21844840 0.45948343 0.4594834
#>  [3,] 0.99783801 0.44573118 0.1555657
#>  [4,] 0.50825615 0.09387907 0.1135642
#>  [5,] 0.03015751 0.61730603 0.7046965
#>  [6,] 0.01105008 0.90322877 0.2633285
#>  [7,] 0.36293156 0.36293156 0.3153829
#>  [8,] 0.40406882 0.31831790 0.1487895
#>  [9,] 0.40681159 0.61841948 0.4068116
#> [10,] 0.12695091 0.12695091 0.1269509

## Gamma
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma")
#>              [,1]        [,2]        [,3]
#>  [1,] 0.029608356 0.195110207 0.195110207
#>  [2,] 0.098446123 0.098446123 0.098446123
#>  [3,] 0.066756398 1.053936144 0.066756398
#>  [4,] 0.115286466 0.012913724 0.916432027
#>  [5,] 0.193094311 0.193094311 0.193094311
#>  [6,] 0.063669971 1.069598535 1.371628060
#>  [7,] 0.001255647 0.001255647 0.001255647
#>  [8,] 0.700208059 0.075219264 0.481751263
#>  [9,] 0.253708247 0.409027760 0.353140415
#> [10,] 0.378528155 0.702440742 0.315247709
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "ESM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.23530660 0.148788164 0.19684281
#>  [2,] 0.60320346 0.095309065 0.09530907
#>  [3,] 0.19522716 0.195227155 0.19522716
#>  [4,] 0.34939452 0.349394523 0.34939452
#>  [5,] 0.01605419 0.016054186 0.01605419
#>  [6,] 0.34497491 0.772879473 0.77287947
#>  [7,] 0.70824422 0.708244218 0.30093481
#>  [8,] 0.38981553 0.389815531 0.06461228
#>  [9,] 0.26410937 0.420334569 0.42033457
#> [10,] 0.11304945 0.009899398 0.24976517
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "MDCM")
#>             [,1]        [,2]       [,3]
#>  [1,] 0.17009304 0.170093036 0.17009304
#>  [2,] 0.33479289 0.338129167 0.33479289
#>  [3,] 0.18342725 0.134919489 0.29948425
#>  [4,] 0.45273412 0.365076597 0.45273412
#>  [5,] 0.15483690 0.178710918 0.17871092
#>  [6,] 0.04535360 0.067720244 0.04535360
#>  [7,] 0.05950168 0.593896299 0.69747457
#>  [8,] 0.49946931 0.001693142 0.01362152
#>  [9,] 0.15273270 0.670103271 0.39463015
#> [10,] 0.10494132 0.574872444 0.51184135
rpextmo(10, 3, a = 0.2, b = 0.5, gamma = 2, eta = 0.5, family = "Gamma", method = "AM")
#>             [,1]        [,2]        [,3]
#>  [1,] 0.42013695 0.106912968 0.420136952
#>  [2,] 0.03384911 0.598497098 0.033849108
#>  [3,] 0.12372319 0.248199124 0.194244982
#>  [4,] 0.05148995 0.005841248 0.338753393
#>  [5,] 0.78619612 0.515372399 0.005130478
#>  [6,] 0.70137098 1.567108656 0.106565003
#>  [7,] 0.83779768 0.036224262 2.064647429
#>  [8,] 0.11026794 0.110267938 0.110267938
#>  [9,] 0.26126813 1.946595377 0.355623464
#> [10,] 0.39714426 0.213012480 0.746490699